\alias{atkActionSetDescription}
\name{atkActionSetDescription}
\title{atkActionSetDescription}
\description{Sets a description of the specified action of the object.}
\usage{atkActionSetDescription(object, i, desc)}
\arguments{
\item{\code{object}}{[\code{\link{AtkAction}}]  a \code{\link{GObject}} instance that implements AtkActionIface}
\item{\code{i}}{[integer]  the action index corresponding to the action to be performed }
\item{\code{desc}}{[character]  the description to be assigned to this action}
}
\value{[logical]  a gboolean representing if the description was successfully set;}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
