\alias{atkAddFocusTracker}
\name{atkAddFocusTracker}
\title{atkAddFocusTracker}
\description{Adds the specified function to the list of functions to be called
when an object receives focus.}
\usage{atkAddFocusTracker(focus.tracker)}
\arguments{\item{\verb{focus.tracker}}{[AtkEventListener]  Function to be added to the list of functions to be called
when an object receives focus.}}
\value{[numeric]  added focus tracker id, or 0 on failure.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
