\alias{atkComponentGetAlpha}
\name{atkComponentGetAlpha}
\title{atkComponentGetAlpha}
\description{Returns the alpha value (i.e. the opacity) for this
\code{component}, on a scale from 0 (fully transparent) to 1.0
(fully opaque).}
\usage{atkComponentGetAlpha(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkComponent}}]  an \code{\link{AtkComponent}}}}
\details{  Since 1.12}
\value{[numeric]  An alpha value from 0 to 1.0, inclusive.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
