\alias{atkComponentGetSize}
\name{atkComponentGetSize}
\title{atkComponentGetSize}
\description{Gets the size of the \code{component} in terms of width and height.}
\usage{atkComponentGetSize(object)}
\arguments{\item{\verb{object}}{[\verb{\link{AtkComponent}}]  an \verb{\link{AtkComponent}}}}
\value{
A list containing the following elements:
\item{\verb{width}}{[integer]  \verb{integer} to put width of \code{component}}
\item{\verb{height}}{[integer]  \verb{integer} to put height of \code{component}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
