\alias{atkComponentRemoveFocusHandler}
\name{atkComponentRemoveFocusHandler}
\title{atkComponentRemoveFocusHandler}
\description{Remove the handler specified by \code{handler.id} from the list of
functions to be executed when this object receives focus events 
(in or out).}
\usage{atkComponentRemoveFocusHandler(object, handler.id)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkComponent}}]  the \code{\link{AtkComponent}} to remove the focus handler from}
\item{\verb{handler.id}}{[numeric]  the handler id of the focus handler to be removed
from \code{component}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
