\alias{atkDocumentGetAttributes}
\name{atkDocumentGetAttributes}
\title{atkDocumentGetAttributes}
\description{Gets an AtkAttributeSet which describes document-wide
         attributes as name-value pairs.}
\usage{atkDocumentGetAttributes(object)}
\arguments{\item{\verb{object}}{[\verb{\link{AtkDocument}}]  a \verb{\link{GObject}} instance that implements AtkDocumentIface}}
\details{  Since  ATK 1.12}
\value{[\verb{\link{AtkAttributeSet}}]  An AtkAttributeSet containing the explicitly
         set name-value-pair attributes associated with this document
         as a whole.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
