\alias{atkEditableTextCutText}
\name{atkEditableTextCutText}
\title{atkEditableTextCutText}
\description{Copy text from \code{start.pos} up to, but not including \code{end.pos}
to the clipboard and then delete from the widget.}
\usage{atkEditableTextCutText(object, start.pos, end.pos)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkEditableText}}]  an \code{\link{AtkEditableText}}}
\item{\verb{start.pos}}{[integer]  start position}
\item{\verb{end.pos}}{[integer]  end position}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
