\alias{atkEditableTextPasteText}
\name{atkEditableTextPasteText}
\title{atkEditableTextPasteText}
\description{Paste text from clipboard to specified \code{position}.}
\usage{atkEditableTextPasteText(object, position)}
\arguments{
\item{\verb{object}}{[\verb{\link{AtkEditableText}}]  an \verb{\link{AtkEditableText}}}
\item{\verb{position}}{[integer]  position to paste}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
