\alias{atkFocusTrackerInit}
\name{atkFocusTrackerInit}
\title{atkFocusTrackerInit}
\description{Specifies the function to be called for focus tracker initialization.
This function should be called by an implementation of the
ATK interface if any specific work needs to be done to enable
focus tracking.}
\usage{atkFocusTrackerInit(add.function)}
\arguments{\item{\verb{add.function}}{[AtkEventListenerInit]  Function to be called for focus tracker initialization}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
