\alias{atkFocusTrackerNotify}
\name{atkFocusTrackerNotify}
\title{atkFocusTrackerNotify}
\description{Cause the focus tracker functions which have been specified to be
executed for the object.}
\usage{atkFocusTrackerNotify(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkObject}}]  an \code{\link{AtkObject}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
