\alias{atkGetFocusObject}
\name{atkGetFocusObject}
\title{atkGetFocusObject}
\description{Gets the currently focused object.}
\usage{atkGetFocusObject()}
\value{[\code{\link{AtkObject}}]  the currently focused object for the current application}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
