\alias{atkGetRoot}
\name{atkGetRoot}
\title{atkGetRoot}
\description{Gets the root accessible container for the current application.}
\usage{atkGetRoot()}
\value{[\code{\link{AtkObject}}]  the root accessible container for the current application}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
