\alias{atkGetToolkitName}
\name{atkGetToolkitName}
\title{atkGetToolkitName}
\description{Gets name string for the GUI toolkit implementing ATK for this application.}
\usage{atkGetToolkitName()}
\value{[character]  name string for the GUI toolkit implementing ATK for this application}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
