\alias{atkGetToolkitVersion}
\name{atkGetToolkitVersion}
\title{atkGetToolkitVersion}
\description{Gets version string for the GUI toolkit implementing ATK for this application.}
\usage{atkGetToolkitVersion()}
\value{[character]  version string for the GUI toolkit implementing ATK for this application}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
