\alias{atkHyperlinkGetEndIndex}
\name{atkHyperlinkGetEndIndex}
\title{atkHyperlinkGetEndIndex}
\description{Gets the index with the hypertext document at which this link ends.}
\usage{atkHyperlinkGetEndIndex(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkHyperlink}}]  an \code{\link{AtkHyperlink}}}}
\value{[integer]  the index with the hypertext document at which this link ends}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
