\alias{atkHyperlinkGetStartIndex}
\name{atkHyperlinkGetStartIndex}
\title{atkHyperlinkGetStartIndex}
\description{Gets the index with the hypertext document at which this link begins.}
\usage{atkHyperlinkGetStartIndex(object)}
\arguments{\item{\code{object}}{[\code{\link{AtkHyperlink}}]  an \code{\link{AtkHyperlink}}}}
\value{[integer]  the index with the hypertext document at which this link begins}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
