\alias{atkHyperlinkImplGetHyperlink}
\name{atkHyperlinkImplGetHyperlink}
\title{atkHyperlinkImplGetHyperlink}
\description{Gets the hyperlink associated with this object.}
\usage{atkHyperlinkImplGetHyperlink(object)}
\arguments{\item{\code{object}}{[\code{\link{AtkHyperlinkImpl}}]  a GObject instance that implements AtkHyperlinkImplIface}}
\details{  Since  ATK 1.12}
\value{[\code{\link{AtkHyperlink}}] an AtkHyperlink object which points to this implementing AtkObject.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
