\alias{atkHyperlinkIsSelectedLink}
\name{atkHyperlinkIsSelectedLink}
\title{atkHyperlinkIsSelectedLink}
\description{Determines whether this AtkHyperlink is selected}
\usage{atkHyperlinkIsSelectedLink(object)}
\arguments{\item{\code{object}}{[\code{\link{AtkHyperlink}}]  an \code{\link{AtkHyperlink}}}}
\value{[logical]  True is the AtkHyperlink is selected, False otherwise}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
