\alias{atkHyperlinkIsValid}
\name{atkHyperlinkIsValid}
\title{atkHyperlinkIsValid}
\description{Since the document that a link is associated with may have changed
this method returns \code{TRUE} if the link is still valid (with
respect to the document it references) and \code{FALSE} otherwise.}
\usage{atkHyperlinkIsValid(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkHyperlink}}]  an \code{\link{AtkHyperlink}}}}
\value{[logical]  whether or not this link is still valid}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
