\alias{atkHypertextGetLink}
\name{atkHypertextGetLink}
\title{atkHypertextGetLink}
\description{Gets the link in this hypertext document at index 
\code{link.index}}
\usage{atkHypertextGetLink(object, link.index)}
\arguments{
\item{\code{object}}{[\code{\link{AtkHypertext}}]  an \code{\link{AtkHypertext}}}
\item{\code{link.index}}{[integer]  an integer specifying the desired link}
}
\value{[\code{\link{AtkHyperlink}}]  the link in this hypertext document at
index \code{link.index}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
