\alias{atkHypertextGetNLinks}
\name{atkHypertextGetNLinks}
\title{atkHypertextGetNLinks}
\description{Gets the number of links within this hypertext document.}
\usage{atkHypertextGetNLinks(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkHypertext}}]  an \code{\link{AtkHypertext}}}}
\value{[integer]  the number of links within this hypertext document}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
