\alias{atkImageGetImageLocale}
\name{atkImageGetImageLocale}
\title{atkImageGetImageLocale}
\description{Since ATK 1.12}
\usage{atkImageGetImageLocale(object)}
\arguments{\item{\code{object}}{[\code{\link{AtkImage}}]  An \code{\link{AtkImage}}}}
\value{[character] a string corresponding to the POSIX LC\_MESSAGES locale used by the image description, or NULL if the image does not specify a locale. }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
