\alias{atkNoOpObjectFactoryNew}
\name{atkNoOpObjectFactoryNew}
\title{atkNoOpObjectFactoryNew}
\description{Creates an instance of an \verb{\link{AtkObjectFactory}} which generates primitive
(non-functioning) \verb{AtkObjects}.}
\usage{atkNoOpObjectFactoryNew()}
\value{[\verb{\link{AtkObjectFactory}}]  an instance of an \verb{\link{AtkObjectFactory}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
