\alias{atkNoOpObjectNew}
\name{atkNoOpObjectNew}
\title{atkNoOpObjectNew}
\description{Provides a default (non-functioning stub) \verb{\link{AtkObject}}.
Application maintainers should not use this method.}
\usage{atkNoOpObjectNew(obj)}
\arguments{\item{\verb{obj}}{[\verb{\link{GObject}}]  a \verb{\link{GObject}}}}
\value{[\verb{\link{AtkObject}}]  a default (non-functioning stub) \verb{\link{AtkObject}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
