\alias{atkObjectFactoryCreateAccessible}
\name{atkObjectFactoryCreateAccessible}
\title{atkObjectFactoryCreateAccessible}
\description{Provides an \verb{\link{AtkObject}} that implements an accessibility interface 
on behalf of \code{obj}}
\usage{atkObjectFactoryCreateAccessible(object, obj)}
\arguments{
\item{\verb{object}}{[\verb{\link{AtkObjectFactory}}]  The \verb{\link{AtkObjectFactory}} associated with \code{obj}'s
object type}
\item{\verb{obj}}{[\verb{\link{GObject}}]  a \verb{\link{GObject}} }
}
\value{[\verb{\link{AtkObject}}]  an \verb{\link{AtkObject}} that implements an accessibility interface
on behalf of \code{obj}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
