\alias{atkObjectGetDescription}
\name{atkObjectGetDescription}
\title{atkObjectGetDescription}
\description{Gets the accessible description of the accessible.}
\usage{atkObjectGetDescription(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkObject}}]  an \code{\link{AtkObject}}}}
\value{[character]  a character string representing the accessible description
of the accessible.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
