\alias{atkObjectGetIndexInParent}
\name{atkObjectGetIndexInParent}
\title{atkObjectGetIndexInParent}
\description{Gets the 0-based index of this accessible in its parent; returns -1 if the
accessible does not have an accessible parent.}
\usage{atkObjectGetIndexInParent(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkObject}}]  an \code{\link{AtkObject}}}}
\value{[integer]  an integer which is the index of the accessible in its parent}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
