\alias{atkObjectGetLayer}
\name{atkObjectGetLayer}
\title{atkObjectGetLayer}
\description{Gets the layer of the accessible.}
\usage{atkObjectGetLayer(object)}
\arguments{\item{\code{object}}{[\code{\link{AtkObject}}]  an \code{\link{AtkObject}}}}
\value{[\code{\link{AtkLayer}}]  an \code{\link{AtkLayer}} which is the layer of the accessible

\code{Deprecated}: Use atk\_component\_get\_layer instead.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
