\alias{atkObjectGetNAccessibleChildren}
\name{atkObjectGetNAccessibleChildren}
\title{atkObjectGetNAccessibleChildren}
\description{Gets the number of accessible children of the accessible.}
\usage{atkObjectGetNAccessibleChildren(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkObject}}]  an \code{\link{AtkObject}}}}
\value{[integer]  an integer representing the number of accessible children
of the accessible.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
