\alias{atkObjectGetName}
\name{atkObjectGetName}
\title{atkObjectGetName}
\description{Gets the accessible name of the accessible.}
\usage{atkObjectGetName(object)}
\arguments{\item{\verb{object}}{[\verb{\link{AtkObject}}]  an \verb{\link{AtkObject}}}}
\value{[character]  a character string representing the accessible name of the object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
