\alias{atkObjectGetParent}
\name{atkObjectGetParent}
\title{atkObjectGetParent}
\description{Gets the accessible parent of the accessible.}
\usage{atkObjectGetParent(object)}
\arguments{\item{\verb{object}}{[\verb{\link{AtkObject}}]  an \verb{\link{AtkObject}}}}
\value{[\verb{\link{AtkObject}}]  a \verb{\link{AtkObject}} representing the accessible parent of the accessible}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
