\alias{atkObjectRefRelationSet}
\name{atkObjectRefRelationSet}
\title{atkObjectRefRelationSet}
\description{Gets the \verb{\link{AtkRelationSet}} associated with the object.}
\usage{atkObjectRefRelationSet(object)}
\arguments{\item{\verb{object}}{[\verb{\link{AtkObject}}]  an \verb{\link{AtkObject}}}}
\value{[\verb{\link{AtkRelationSet}}]  an \verb{\link{AtkRelationSet}} representing the relation set of the object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
