\alias{atkObjectRefStateSet}
\name{atkObjectRefStateSet}
\title{atkObjectRefStateSet}
\description{Gets a reference to the state set of the accessible; the caller must
unreference it when it is no longer needed.}
\usage{atkObjectRefStateSet(object)}
\arguments{\item{\verb{object}}{[\verb{\link{AtkObject}}]  an \verb{\link{AtkObject}}}}
\value{[\verb{\link{AtkStateSet}}]  a reference to an \verb{\link{AtkStateSet}} which is the state
set of the accessible}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
