\alias{atkObjectRemovePropertyChangeHandler}
\name{atkObjectRemovePropertyChangeHandler}
\title{atkObjectRemovePropertyChangeHandler}
\description{Removes a property change handler.}
\usage{atkObjectRemovePropertyChangeHandler(object, handler.id)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkObject}}]  an \code{\link{AtkObject}}}
\item{\verb{handler.id}}{[numeric]  a guint which identifies the handler to be removed.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
