\alias{atkObjectSetDescription}
\name{atkObjectSetDescription}
\title{atkObjectSetDescription}
\description{Sets the accessible description of the accessible.}
\usage{atkObjectSetDescription(object, description)}
\arguments{
\item{\code{object}}{[\code{\link{AtkObject}}]  an \code{\link{AtkObject}}}
\item{\code{description}}{[character]  a character string to be set as the accessible description}
}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
