\alias{atkRelationGetRelationType}
\name{atkRelationGetRelationType}
\title{atkRelationGetRelationType}
\description{Gets the type of \code{relation}}
\usage{atkRelationGetRelationType(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkRelation}}]  an \code{\link{AtkRelation}} }}
\value{[\code{\link{AtkRelationType}}]  the type of \code{relation}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
