\alias{atkRelationGetTarget}
\name{atkRelationGetTarget}
\title{atkRelationGetTarget}
\description{Gets the target list of \code{relation}}
\usage{atkRelationGetTarget(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkRelation}}]  an \code{\link{AtkRelation}}}}
\value{[GPtrArray]  the target list of \code{relation}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
