\alias{atkRelationSetGetNRelations}
\name{atkRelationSetGetNRelations}
\title{atkRelationSetGetNRelations}
\description{Determines the number of relations in a relation set.}
\usage{atkRelationSetGetNRelations(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkRelationSet}}]  an \code{\link{AtkRelationSet}}}}
\value{[integer]  an integer representing the number of relations in the set.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
