\alias{atkRelationSetGetRelationByType}
\name{atkRelationSetGetRelationByType}
\title{atkRelationSetGetRelationByType}
\description{Finds a relation that matches the specified type.}
\usage{atkRelationSetGetRelationByType(object, relationship)}
\arguments{
\item{\verb{object}}{[\verb{\link{AtkRelationSet}}]  an \verb{\link{AtkRelationSet}}}
\item{\verb{relationship}}{[\code{\link{AtkRelationType}}]  an \code{\link{AtkRelationType}}}
}
\value{[\verb{\link{AtkRelation}}]  an \verb{\link{AtkRelation}}, which is a relation matching the specified type.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
