\alias{atkRelationSetNew}
\name{atkRelationSetNew}
\title{atkRelationSetNew}
\description{Creates a new empty relation set.}
\usage{atkRelationSetNew()}
\value{[\verb{\link{AtkRelationSet}}]  a new \verb{\link{AtkRelationSet}} }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
