\alias{atkRelationTypeGetName}
\name{atkRelationTypeGetName}
\title{atkRelationTypeGetName}
\description{Gets the description string describing the \code{\link{AtkRelationType}} \code{type}.}
\usage{atkRelationTypeGetName(type)}
\arguments{\item{\verb{type}}{[\code{\link{AtkRelationType}}]  The \code{\link{AtkRelationType}} whose name is required}}
\value{[character]  the string describing the AtkRelationType}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
