\alias{atkRelationTypeRegister}
\name{atkRelationTypeRegister}
\title{atkRelationTypeRegister}
\description{Associate \code{name} with a new \code{\link{AtkRelationType}}}
\usage{atkRelationTypeRegister(name)}
\arguments{\item{\verb{name}}{[character]  a name string}}
\value{[\code{\link{AtkRelationType}}]  an \code{\link{AtkRelationType}} associated with \code{name}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
