\alias{atkRemoveGlobalEventListener}
\name{atkRemoveGlobalEventListener}
\title{atkRemoveGlobalEventListener}
\description{Removes the specified event listener}
\usage{atkRemoveGlobalEventListener(listener.id)}
\arguments{\item{\verb{listener.id}}{[numeric]  the id of the event listener to remove}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
