\alias{atkRoleGetLocalizedName}
\name{atkRoleGetLocalizedName}
\title{atkRoleGetLocalizedName}
\description{Gets the localized description string describing the \code{\link{AtkRole}} \code{role}.}
\usage{atkRoleGetLocalizedName(role)}
\arguments{\item{\verb{role}}{[\code{\link{AtkRole}}]  The \code{\link{AtkRole}} whose localized name is required}}
\value{[character]  the localized string describing the AtkRole}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
