\alias{atkRoleGetName}
\name{atkRoleGetName}
\title{atkRoleGetName}
\description{Gets the description string describing the \code{\link{AtkRole}} \code{role}.}
\usage{atkRoleGetName(role)}
\arguments{\item{\verb{role}}{[\code{\link{AtkRole}}]  The \code{\link{AtkRole}} whose name is required}}
\value{[character]  the string describing the AtkRole}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
