\alias{atkSelectionSelectAllSelection}
\name{atkSelectionSelectAllSelection}
\title{atkSelectionSelectAllSelection}
\description{Causes every child of the object to be selected if the object
supports multiple selections.}
\usage{atkSelectionSelectAllSelection(object)}
\arguments{\item{\code{object}}{[\code{\link{AtkSelection}}]  a \code{\link{GObject}} instance that implements AtkSelectionIface}}
\value{[logical]  TRUE if success, FALSE otherwise.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
