\alias{atkStateSetAndSets}
\name{atkStateSetAndSets}
\title{atkStateSetAndSets}
\description{Constructs the intersection of the two sets, returning \code{NULL} if the
intersection is empty.}
\usage{atkStateSetAndSets(object, compare.set)}
\arguments{
\item{\verb{object}}{[\verb{\link{AtkStateSet}}]  an \verb{\link{AtkStateSet}}}
\item{\verb{compare.set}}{[\verb{\link{AtkStateSet}}]  another \verb{\link{AtkStateSet}}}
}
\value{[\verb{\link{AtkStateSet}}]  a new \verb{\link{AtkStateSet}} which is the intersection of the two sets.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
