\alias{atkStateSetContainsState}
\name{atkStateSetContainsState}
\title{atkStateSetContainsState}
\description{Checks whether the state for the specified type is in the specified set.}
\usage{atkStateSetContainsState(object, type)}
\arguments{
\item{\code{object}}{[\code{\link{AtkStateSet}}]  an \code{\link{AtkStateSet}}}
\item{\code{type}}{[\code{\link{AtkStateType}}]  an \code{\link{AtkStateType}}}
}
\value{[logical]  \code{TRUE} if \code{type} is the state type is in \code{set}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
