\alias{atkStateSetIsEmpty}
\name{atkStateSetIsEmpty}
\title{atkStateSetIsEmpty}
\description{Checks whether the state set is empty, i.e. has no states set.}
\usage{atkStateSetIsEmpty(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkStateSet}}]  an \code{\link{AtkStateType}}}}
\value{[logical] \code{TRUE} if \code{set} has no states set, otherwise \code{FALSE}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
