\alias{atkStateSetNew}
\name{atkStateSetNew}
\title{atkStateSetNew}
\description{Creates a new empty state set.}
\usage{atkStateSetNew()}
\value{[\code{\link{AtkStateSet}}]  a new \code{\link{AtkStateSet}} }
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
