\alias{atkStateSetOrSets}
\name{atkStateSetOrSets}
\title{atkStateSetOrSets}
\description{Constructs the union of the two sets.}
\usage{atkStateSetOrSets(object, compare.set)}
\arguments{
\item{\verb{object}}{[\verb{\link{AtkStateSet}}]  an \verb{\link{AtkStateSet}}}
\item{\verb{compare.set}}{[\verb{\link{AtkStateSet}}]  another \verb{\link{AtkStateSet}}}
}
\value{[\verb{\link{AtkStateSet}}]  a new \verb{\link{AtkStateSet}} which is the union of the two sets,
returning \code{NULL} is empty.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
