\alias{atkStateSetRemoveState}
\name{atkStateSetRemoveState}
\title{atkStateSetRemoveState}
\description{Removes the state for the specified type from the state set.}
\usage{atkStateSetRemoveState(object, type)}
\arguments{
\item{\code{object}}{[\code{\link{AtkStateSet}}]  an \code{\link{AtkStateSet}}}
\item{\code{type}}{[\code{\link{AtkStateType}}]  an \code{AtkType}}
}
\value{[logical]  \code{TRUE} if \code{type} was the state type is in \code{set}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
