\alias{atkStateTypeForName}
\name{atkStateTypeForName}
\title{atkStateTypeForName}
\description{Gets the \code{\link{AtkStateType}} corresponding to the description string \code{name}.}
\usage{atkStateTypeForName(name)}
\arguments{\item{\verb{name}}{[character]  a character string state name}}
\value{[\code{\link{AtkStateType}}]  an \code{\link{AtkStateType}} corresponding to \code{name} }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
