\alias{atkStateTypeGetName}
\name{atkStateTypeGetName}
\title{atkStateTypeGetName}
\description{Gets the description string describing the \code{\link{AtkStateType}} \code{type}.}
\usage{atkStateTypeGetName(type)}
\arguments{\item{\verb{type}}{[\code{\link{AtkStateType}}]  The \code{\link{AtkStateType}} whose name is required}}
\value{[character]  the string describing the AtkStateType}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
