\alias{atkStateTypeRegister}
\name{atkStateTypeRegister}
\title{atkStateTypeRegister}
\description{Register a new object state.}
\usage{atkStateTypeRegister(name)}
\arguments{\item{\code{name}}{[character]  a character string describing the new state.}}
\value{[\code{\link{AtkStateType}}]  an \code{numeric} value for the new state.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
