\alias{atkStreamableContentGetNMimeTypes}
\name{atkStreamableContentGetNMimeTypes}
\title{atkStreamableContentGetNMimeTypes}
\description{Gets the number of mime types supported by this object.}
\usage{atkStreamableContentGetNMimeTypes(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkStreamableContent}}]  a GObject instance that implements AtkStreamableContentIface}}
\value{[integer]  a gint which is the number of mime types supported by the object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
