\alias{atkTableGetCaption}
\name{atkTableGetCaption}
\title{atkTableGetCaption}
\description{Gets the caption for the \code{table}.}
\usage{atkTableGetCaption(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkTable}}]  a GObject instance that implements AtkTableInterface}}
\value{[\code{\link{AtkObject}}]  a AtkObject* representing the table caption, or \code{NULL}
if value does not implement this interface.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
