\alias{atkTableGetNColumns}
\name{atkTableGetNColumns}
\title{atkTableGetNColumns}
\description{Gets the number of columns in the table.}
\usage{atkTableGetNColumns(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkTable}}]  a GObject instance that implements AtkTableIface}}
\value{[integer]  a gint representing the number of columns, or 0
if value does not implement this interface.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
