\alias{atkTableGetNRows}
\name{atkTableGetNRows}
\title{atkTableGetNRows}
\description{Gets the number of rows in the table.}
\usage{atkTableGetNRows(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkTable}}]  a GObject instance that implements AtkTableIface}}
\value{[integer]  a gint representing the number of rows, or 0
if value does not implement this interface.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
