\alias{atkTableGetSummary}
\name{atkTableGetSummary}
\title{atkTableGetSummary}
\description{Gets the summary description of the table.}
\usage{atkTableGetSummary(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkTable}}]  a GObject instance that implements AtkTableIface}}
\value{[\code{\link{AtkObject}}]  a AtkObject* representing a summary description of the table,
or zero if value does not implement this interface.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
