\alias{atkTableSetColumnDescription}
\name{atkTableSetColumnDescription}
\title{atkTableSetColumnDescription}
\description{Sets the description text for the specified \code{column} of the \code{table}.}
\usage{atkTableSetColumnDescription(object, column, description)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkTable}}]  a GObject instance that implements AtkTableIface}
\item{\verb{column}}{[integer]  a \verb{integer} representing a column in \code{table}}
\item{\verb{description}}{[character]  a \verb{character} representing the description text
to set for the specified \code{column} of the \code{table}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
