\alias{atkTableSetRowDescription}
\name{atkTableSetRowDescription}
\title{atkTableSetRowDescription}
\description{Sets the description text for the specified \code{row} of \code{table}.}
\usage{atkTableSetRowDescription(object, row, description)}
\arguments{
\item{\code{object}}{[\code{\link{AtkTable}}]  a GObject instance that implements AtkTableIface}
\item{\code{row}}{[integer]  a \code{integer} representing a row in \code{table}}
\item{\code{description}}{[character]  a \code{character} representing the description text
to set for the specified \code{row} of \code{table}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
