\alias{atkTextAttributeGetName}
\name{atkTextAttributeGetName}
\title{atkTextAttributeGetName}
\description{Gets the name corresponding to the \code{\link{AtkTextAttribute}}}
\usage{atkTextAttributeGetName(attr)}
\arguments{\item{\code{attr}}{[\code{\link{AtkTextAttribute}}]  The \code{\link{AtkTextAttribute}} whose name is required}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
