\alias{atkTextAttributeRegister}
\name{atkTextAttributeRegister}
\title{atkTextAttributeRegister}
\description{Associate \code{name} with a new \code{\link{AtkTextAttribute}}}
\usage{atkTextAttributeRegister(name)}
\arguments{\item{\verb{name}}{[character]  a name string}}
\value{[\code{\link{AtkTextAttribute}}]  an \code{\link{AtkTextAttribute}} associated with \code{name}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
