\alias{atkTextGetCaretOffset}
\name{atkTextGetCaretOffset}
\title{atkTextGetCaretOffset}
\description{Gets the offset position of the caret (cursor).}
\usage{atkTextGetCaretOffset(object)}
\arguments{\item{\verb{object}}{[\verb{\link{AtkText}}]  an \verb{\link{AtkText}}}}
\value{[integer]  the offset position of the caret (cursor).}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
