\alias{atkTextGetCharacterCount}
\name{atkTextGetCharacterCount}
\title{atkTextGetCharacterCount}
\description{Gets the character count.}
\usage{atkTextGetCharacterCount(object)}
\arguments{\item{\verb{object}}{[\verb{\link{AtkText}}]  an \verb{\link{AtkText}}}}
\value{[integer]  the number of characters.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
