\alias{atkTextGetDefaultAttributes}
\name{atkTextGetDefaultAttributes}
\title{atkTextGetDefaultAttributes}
\description{Creates an \verb{\link{AtkAttributeSet}} which consists of the default values of
attributes for the text. See the enum AtkTextAttribute for types of text 
attributes that can be returned. Note that other attributes may also be 
returned.}
\usage{atkTextGetDefaultAttributes(object)}
\arguments{\item{\verb{object}}{[\verb{\link{AtkText}}]  an \verb{\link{AtkText}}}}
\value{[\verb{\link{AtkAttributeSet}}]  an \verb{\link{AtkAttributeSet}} which contains the default values of attributes.
at \code{offset}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
