\alias{atkTextGetNSelections}
\name{atkTextGetNSelections}
\title{atkTextGetNSelections}
\description{Gets the number of selected regions.}
\usage{atkTextGetNSelections(object)}
\arguments{\item{\code{object}}{[\code{\link{AtkText}}]  an \code{\link{AtkText}}}}
\value{[integer]  The number of selected regions, or -1 if a failure
  occurred.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
