\alias{atkTextSetCaretOffset}
\name{atkTextSetCaretOffset}
\title{atkTextSetCaretOffset}
\description{Sets the caret (cursor) position to the specified \code{offset}.}
\usage{atkTextSetCaretOffset(object, offset)}
\arguments{
\item{\code{object}}{[\code{\link{AtkText}}]  an \code{\link{AtkText}}}
\item{\code{offset}}{[integer]  position}
}
\value{[logical]  \code{TRUE} if success, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
