\alias{atkValueGetMaximumValue}
\name{atkValueGetMaximumValue}
\title{atkValueGetMaximumValue}
\description{Gets the maximum value of this object.}
\usage{atkValueGetMaximumValue(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkValue}}]  a GObject instance that implements AtkValueIface}}
\value{
A list containing the following elements:
\item{\verb{value}}{[R object]  a \verb{R object} representing the maximum accessible value}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
