\alias{atkValueGetMinimumValue}
\name{atkValueGetMinimumValue}
\title{atkValueGetMinimumValue}
\description{Gets the minimum value of this object.}
\usage{atkValueGetMinimumValue(object)}
\arguments{\item{\code{object}}{[\code{\link{AtkValue}}]  a GObject instance that implements AtkValueIface}}
\value{
A list containing the following elements:
\item{\code{value}}{[R object]  a \code{R object} representing the minimum accessible value}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
