\alias{atkValueSetCurrentValue}
\name{atkValueSetCurrentValue}
\title{atkValueSetCurrentValue}
\description{Sets the value of this object.}
\usage{atkValueSetCurrentValue(object, value)}
\arguments{
\item{\code{object}}{[\code{\link{AtkValue}}]  a GObject instance that implements AtkValueIface}
\item{\code{value}}{[R object]  a \code{R object} which is the desired new accessible value.}
}
\value{[logical] \code{TRUE} if new value is successfully set, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
