\alias{cairo-PDF-Surfaces}
\name{cairo-PDF-Surfaces}
\title{PDF Surfaces}
\description{Rendering PDF documents}
\section{Methods and Functions}{
\code{\link{cairoPdfSurfaceCreate}(filename, width.in.points, height.in.points)}\cr
\code{\link{cairoPdfSurfaceCreateForStream}(write.func, closure, width.in.points, height.in.points)}\cr
\code{\link{cairoPdfSurfaceSetSize}(surface, width.in.points, height.in.points)}\cr
}
\references{\url{http://www.cairographics.org/manual/cairo-PDF-Surfaces.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
