\alias{cairo-PostScript-Surfaces}
\name{cairo-PostScript-Surfaces}
\title{PostScript Surfaces}
\description{Rendering PostScript documents}
\section{Methods and Functions}{
\code{\link{cairoPsSurfaceCreate}(filename, width.in.points, height.in.points)}\cr
\code{\link{cairoPsSurfaceCreateForStream}(write.func, closure, width.in.points, height.in.points)}\cr
\code{\link{cairoPsSurfaceSetSize}(surface, width.in.points, height.in.points)}\cr
\code{\link{cairoPsSurfaceDscBeginSetup}(surface)}\cr
\code{\link{cairoPsSurfaceDscBeginPageSetup}(surface)}\cr
\code{\link{cairoPsSurfaceDscComment}(surface, comment)}\cr
}
\references{\url{http://www.cairographics.org/manual/cairo-PostScript-Surfaces.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
