\alias{cairo-quartz-font}
\name{cairo-quartz-font}
\title{Quartz (CGFont) Fonts}
\description{Font support via CGFont on OS X}
\section{Detailed Description}{The Quartz font backend is primarily used to render text on Apple
MacOS X systems.  The CGFont API is used for the internal
implementation of the font backend methods.}
\references{\url{https://www.cairographics.org/manual/cairo-Quartz-(CGFont)-Fonts.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
