\alias{cairo-Transformations}
\name{cairo-Transformations}
\title{Transformations}
\description{Manipulating the current transformation matrix}
\section{Methods and Functions}{
\code{\link{cairoTranslate}(cr, tx, ty)}\cr
\code{\link{cairoScale}(cr, sx, sy)}\cr
\code{\link{cairoRotate}(cr, angle)}\cr
\code{\link{cairoTransform}(cr, matrix)}\cr
\code{\link{cairoSetMatrix}(cr, matrix)}\cr
\code{\link{cairoGetMatrix}(cr, matrix)}\cr
\code{\link{cairoIdentityMatrix}(cr)}\cr
\code{\link{cairoUserToDevice}(cr, x, y)}\cr
\code{\link{cairoUserToDeviceDistance}(cr, dx, dy)}\cr
\code{\link{cairoDeviceToUser}(cr, x, y)}\cr
\code{\link{cairoDeviceToUserDistance}(cr, dx, dy)}\cr
}
\references{\url{http://www.cairographics.org/manual/cairo-Transformations.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
