\alias{cairo-Version-Information}
\name{cairo-Version-Information}
\title{Version Information}
\description{Compile and run time version checks}
\section{Methods and Functions}{\link{cairoVersionString}()}
\references{\url{http://developer.gnome.org/doc/API/2.0/cairo/cairo-Version-Information.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
