\alias{cairoFontFaceStatus}
\name{cairoFontFaceStatus}
\title{cairoFontFaceStatus}
\description{Checks whether an error has previously occurred for this
font face}
\usage{cairoFontFaceStatus(font.face)}
\arguments{\item{\verb{font.face}}{[\verb{\link{CairoFontFace}}]  a \verb{\link{CairoFontFace}}}}
\value{[\verb{\link{CairoStatus}}] \code{CAIRO_STATUS_SUCCESS} or another error such as
  \code{CAIRO_STATUS_NO_MEMORY}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
