\alias{cairoFontOptionsEqual}
\name{cairoFontOptionsEqual}
\title{cairoFontOptionsEqual}
\description{Compares two font options objects for equality.}
\usage{cairoFontOptionsEqual(options, other)}
\arguments{
\item{\code{options}}{[\code{\link{CairoFontOptions}}]  a \code{\link{CairoFontOptions}}}
\item{\code{other}}{[\code{\link{CairoFontOptions}}]  another \code{\link{CairoFontOptions}}}
}
\value{[logical]  \code{TRUE} if all fields of the two font options objects match}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
