\alias{cairoFontOptionsGetAntialias}
\name{cairoFontOptionsGetAntialias}
\title{cairoFontOptionsGetAntialias}
\description{Gets the antialising mode for the font options object.}
\usage{cairoFontOptionsGetAntialias(options)}
\arguments{\item{\code{options}}{[\code{\link{cairo_font_options_t}}]  a \code{\link{cairo_font_options_t}}}}
\value{[\code{\link{cairo_antialias_t}}]  the antialiasing mode}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
