\alias{cairoFontOptionsGetHintMetrics}
\name{cairoFontOptionsGetHintMetrics}
\title{cairoFontOptionsGetHintMetrics}
\description{Gets the metrics hinting mode for the font options object.
See the documentation for \verb{\link{CairoHintMetrics}} for full details.}
\usage{cairoFontOptionsGetHintMetrics(options)}
\arguments{\item{\verb{options}}{[\verb{\link{CairoFontOptions}}]  a \verb{\link{CairoFontOptions}}}}
\value{[\verb{\link{CairoHintMetrics}}]  the metrics hinting mode for the font options object}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
