\alias{cairoFontOptionsGetHintStyle}
\name{cairoFontOptionsGetHintStyle}
\title{cairoFontOptionsGetHintStyle}
\description{Gets the hint style for font outlines for the font options object.
See the documentation for \code{\link{cairo_hint_style_t}} for full details.}
\usage{cairoFontOptionsGetHintStyle(options)}
\arguments{\item{\code{options}}{[\code{\link{cairo_font_options_t}}]  a \code{\link{cairo_font_options_t}}}}
\value{[\code{\link{cairo_hint_style_t}}]  the hint style for the font options object}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
