\alias{cairoFontOptionsGetSubpixelOrder}
\name{cairoFontOptionsGetSubpixelOrder}
\title{cairoFontOptionsGetSubpixelOrder}
\description{Gets the subpixel order for the font options object.
See the documentation for \code{\link{cairo_subpixel_order_t}} for full details.}
\usage{cairoFontOptionsGetSubpixelOrder(options)}
\arguments{\item{\code{options}}{[\code{\link{cairo_font_options_t}}]  a \code{\link{cairo_font_options_t}}}}
\value{[\code{\link{cairo_subpixel_order_t}}]  the subpixel order for the font options object}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
