\alias{cairoFontOptionsSetHintMetrics}
\name{cairoFontOptionsSetHintMetrics}
\title{cairoFontOptionsSetHintMetrics}
\description{Sets the metrics hinting mode for the font options object. This
controls whether metrics are quantized to integer values in
device units.
See the documentation for \verb{\link{CairoHintMetrics}} for full details.}
\usage{cairoFontOptionsSetHintMetrics(options, hint.metrics)}
\arguments{
\item{\verb{options}}{[\verb{\link{CairoFontOptions}}]  a \verb{\link{CairoFontOptions}}}
\item{\verb{hint.metrics}}{[\verb{\link{CairoHintMetrics}}]  the new metrics hinting mode}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
