\alias{cairoFontOptionsStatus}
\name{cairoFontOptionsStatus}
\title{cairoFontOptionsStatus}
\description{Checks whether an error has previously occurred for this
font options object}
\usage{cairoFontOptionsStatus(options)}
\arguments{\item{\verb{options}}{[\verb{\link{CairoFontOptions}}]  a \verb{\link{CairoFontOptions}}}}
\value{[\verb{\link{CairoStatus}}] \code{CAIRO_STATUS_SUCCESS} or \code{CAIRO_STATUS_NO_MEMORY}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
