\alias{cairoGetAntialias}
\name{cairoGetAntialias}
\title{cairoGetAntialias}
\description{Gets the current shape antialiasing mode, as set by \code{cairoSetShapeAntialias()}.}
\usage{cairoGetAntialias(cr)}
\arguments{\item{\code{cr}}{[\code{\link{Cairo}}]  a cairo context}}
\value{[\code{\link{CairoAntialias}}]  the current shape antialiasing mode.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
