\alias{cairoGetFontMatrix}
\name{cairoGetFontMatrix}
\title{cairoGetFontMatrix}
\description{Stores the current font matrix into \code{matrix}. See
\code{\link{cairoSetFontMatrix}}.}
\usage{cairoGetFontMatrix(cr, matrix)}
\arguments{
\item{\code{cr}}{[\code{\link{cairo_t}}]  a \code{\link{cairo_t}}}
\item{\code{matrix}}{[\code{\link{cairo_matrix_t}}]  return value for the matrix}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
