\alias{cairoGetLineCap}
\name{cairoGetLineCap}
\title{cairoGetLineCap}
\description{Gets the current line cap style, as set by \code{\link{cairoSetLineCap}}.}
\usage{cairoGetLineCap(cr)}
\arguments{\item{\code{cr}}{[\code{\link{cairo_t}}]  a cairo context}}
\value{[\code{\link{cairo_line_cap_t}}]  the current line cap style.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
