\alias{cairoGetLineJoin}
\name{cairoGetLineJoin}
\title{cairoGetLineJoin}
\description{Gets the current line join style, as set by \code{\link{cairoSetLineJoin}}.}
\usage{cairoGetLineJoin(cr)}
\arguments{\item{\code{cr}}{[\code{\link{Cairo}}]  a cairo context}}
\value{[\code{\link{CairoLineJoin}}]  the current line join style.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
