\alias{cairoGetOperator}
\name{cairoGetOperator}
\title{cairoGetOperator}
\description{Gets the current compositing operator for a cairo context.}
\usage{cairoGetOperator(cr)}
\arguments{\item{\verb{cr}}{[\verb{\link{Cairo}}]  a cairo context}}
\value{[\verb{\link{CairoOperator}}]  the current compositing operator.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
