\alias{cairoGetTolerance}
\name{cairoGetTolerance}
\title{cairoGetTolerance}
\description{Gets the current tolerance value, as set by \code{\link{cairoSetTolerance}}.}
\usage{cairoGetTolerance(cr)}
\arguments{\item{\code{cr}}{[\code{\link{Cairo}}]  a cairo context}}
\value{[numeric]  the current tolerance value.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
