\alias{cairoGlyphPath}
\name{cairoGlyphPath}
\title{cairoGlyphPath}
\description{\emph{undocumented
}}
\usage{cairoGlyphPath(cr, glyphs)}
\arguments{
\item{\code{cr}}{[\code{\link{Cairo}}] }
\item{\code{glyphs}}{[\code{\link{CairoGlyph}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
