\alias{cairoHasCurrentPoint}
\name{cairoHasCurrentPoint}
\title{cairoHasCurrentPoint}
\description{Returns whether a current point is defined on the current path.
See \code{\link{cairoGetCurrentPoint}} for details on the current point.}
\usage{cairoHasCurrentPoint(cr)}
\arguments{\item{\verb{cr}}{[\code{\link{Cairo}}]  a cairo context}}
\details{  Since 1.6}
\value{[logical]  whether a current point is defined.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
