\alias{cairoIdentityMatrix}
\name{cairoIdentityMatrix}
\title{cairoIdentityMatrix}
\description{Resets the current transformation matrix (CTM) by setting it equal
to the identity matrix. That is, the user-space and device-space
axes will be aligned and one user-space unit will transform to one
device-space unit.}
\usage{cairoIdentityMatrix(cr)}
\arguments{\item{\verb{cr}}{[\code{\link{Cairo}}]  a cairo context}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
