\alias{cairoImageSurfaceGetHeight}
\name{cairoImageSurfaceGetHeight}
\title{cairoImageSurfaceGetHeight}
\description{Get the height of the image surface in pixels.}
\usage{cairoImageSurfaceGetHeight(surface)}
\arguments{\item{\verb{surface}}{[\code{\link{CairoSurface}}]  a \verb{cairo_image_surface_t}}}
\value{[integer]  the height of the surface in pixels.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
