\alias{cairoImageSurfaceGetWidth}
\name{cairoImageSurfaceGetWidth}
\title{cairoImageSurfaceGetWidth}
\description{Get the width of the image surface in pixels.}
\usage{cairoImageSurfaceGetWidth(surface)}
\arguments{\item{\code{surface}}{[\code{\link{CairoSurface}}]  a \code{cairo_image_surface_t}}}
\value{[integer]  the width of the surface in pixels.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
