\alias{cairoInFill}
\name{cairoInFill}
\title{cairoInFill}
\description{\emph{undocumented
}}
\usage{cairoInFill(cr, x, y)}
\arguments{
\item{\code{cr}}{[\code{\link{Cairo}}] }
\item{\code{x}}{[numeric] }
\item{\code{y}}{[numeric] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
