\alias{cairoInStroke}
\name{cairoInStroke}
\title{cairoInStroke}
\description{\emph{undocumented
}}
\usage{cairoInStroke(cr, x, y)}
\arguments{
\item{\code{cr}}{[\code{\link{Cairo}}] }
\item{\code{x}}{[numeric] }
\item{\code{y}}{[numeric] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
