\alias{cairoMatrixInitIdentity}
\name{cairoMatrixInitIdentity}
\title{cairoMatrixInitIdentity}
\description{Modifies \code{matrix} to be an identity transformation.}
\usage{cairoMatrixInitIdentity()}
\value{
A list containing the following elements:
\item{\code{matrix}}{[\code{\link{cairo_matrix_t}}]  a \code{\link{cairo_matrix_t}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
