\alias{cairoPaint}
\name{cairoPaint}
\title{cairoPaint}
\description{A drawing operator that paints the current source everywhere within
the current clip region.}
\usage{cairoPaint(cr)}
\arguments{\item{\code{cr}}{[\code{\link{Cairo}}]  a cairo context}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
