\alias{cairoPatternGetFilter}
\name{cairoPatternGetFilter}
\title{cairoPatternGetFilter}
\description{Gets the current filter for a pattern.  See \verb{\link{CairoFilter}}
for details on each filter.}
\usage{cairoPatternGetFilter(pattern)}
\arguments{\item{\verb{pattern}}{[\verb{\link{CairoPattern}}]  a \verb{\link{CairoPattern}}}}
\value{[\verb{\link{CairoFilter}}]  the current filter used for resizing the pattern.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
