\alias{cairoPatternGetMatrix}
\name{cairoPatternGetMatrix}
\title{cairoPatternGetMatrix}
\description{Stores the pattern's transformation matrix into \code{matrix}.}
\usage{cairoPatternGetMatrix(pattern, matrix)}
\arguments{
\item{\code{pattern}}{[\code{\link{cairo_pattern_t}}]  a \code{\link{cairo_pattern_t}}}
\item{\code{matrix}}{[\code{\link{cairo_matrix_t}}]  return value for the matrix}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
