\alias{cairoPatternGetType}
\name{cairoPatternGetType}
\title{cairoPatternGetType}
\description{This function returns the type a pattern.
See \verb{\link{CairoPatternType}} for available types.}
\usage{cairoPatternGetType(pattern)}
\arguments{\item{\verb{pattern}}{[\verb{\link{CairoPattern}}]  a \verb{\link{CairoPattern}}}}
\details{  Since  1.2}
\value{[\verb{\link{CairoPatternType}}]  The type of \code{pattern}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
