\alias{cairoPatternSetExtend}
\name{cairoPatternSetExtend}
\title{cairoPatternSetExtend}
\description{Sets the mode to be used for drawing outside the area of a pattern.
See \code{\link{CairoExtend}} for details on the semantics of each extend
strategy.}
\usage{cairoPatternSetExtend(pattern, extend)}
\arguments{
\item{\code{pattern}}{[\code{\link{CairoPattern}}]  a \code{\link{CairoPattern}}}
\item{\code{extend}}{[\code{\link{CairoExtend}}]  a \code{\link{CairoExtend}} describing how the area outside of the
pattern will be drawn}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
