\alias{cairoPatternSetFilter}
\name{cairoPatternSetFilter}
\title{cairoPatternSetFilter}
\usage{cairoPatternSetFilter(pattern, filter)}
\arguments{
\item{\code{pattern}}{[\code{\link{cairo_pattern_t}}] }
\item{\code{filter}}{[\code{\link{cairo_filter_t}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
