\alias{cairoPatternStatus}
\name{cairoPatternStatus}
\title{cairoPatternStatus}
\description{Checks whether an error has previously occurred for this
pattern.}
\usage{cairoPatternStatus(pattern)}
\arguments{\item{\code{pattern}}{[\code{\link{cairo_pattern_t}}]  a \code{\link{cairo_pattern_t}}}}
\value{[\code{\link{cairo_status_t}}]  \code{CAIRO_STATUS_SUCCESS}, \code{CAIRO_STATUS_NO_MEMORY}, or
\code{CAIRO_STATUS_PATTERN_TYPE_MISMATCH}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
