\alias{cairoPsLevelToString}
\name{cairoPsLevelToString}
\title{cairoPsLevelToString}
\description{Get the string representation of the given \code{level} id. This function
will return \code{NULL} if \code{level} id isn't valid. See \code{\link{cairoPsGetLevels}}
for a way to get the list of valid level ids.}
\usage{cairoPsLevelToString(level)}
\arguments{\item{\verb{level}}{[\code{\link{CairoPsLevel}}]  a level id}}
\details{  Since 1.6}
\value{[char]  the string associated to given level.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
