\alias{cairoPsSurfaceGetEps}
\name{cairoPsSurfaceGetEps}
\title{cairoPsSurfaceGetEps}
\description{Check whether the PostScript surface will output Encapsulated PostScript.}
\usage{cairoPsSurfaceGetEps(surface)}
\arguments{\item{\verb{surface}}{[\code{\link{CairoSurface}}]  a PostScript \code{\link{CairoSurface}}}}
\details{  Since 1.6}
\value{[logical] \code{TRUE} if the surface will output Encapsulated PostScript.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
