\alias{cairoScaledFontExtents}
\name{cairoScaledFontExtents}
\title{cairoScaledFontExtents}
\description{Gets the metrics for a \code{\link{cairo_scaled_font_t}}.}
\usage{cairoScaledFontExtents(scaled.font)}
\arguments{\item{\code{scaled.font}}{[\code{\link{cairo_scaled_font_t}}]  a \code{\link{cairo_scaled_font_t}}}}
\value{
A list containing the following elements:
\item{\code{extents}}{[\code{\link{cairo_font_extents_t}}]  a \code{\link{cairo_font_extents_t}} which to store the retrieved extents.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
