\alias{cairoScaledFontGetCtm}
\name{cairoScaledFontGetCtm}
\title{cairoScaledFontGetCtm}
\description{Stores the CTM with which \code{scaled.font} was created into \code{ctm}.}
\usage{cairoScaledFontGetCtm(scaled.font)}
\arguments{\item{\verb{scaled.font}}{[\code{\link{CairoScaledFont}}]  a \code{\link{CairoScaledFont}}}}
\details{  Since 1.2}
\value{
A list containing the following elements:
\item{\verb{ctm}}{[\code{\link{CairoMatrix}}]  return value for the CTM}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
