\alias{cairoScaledFontGetFontFace}
\name{cairoScaledFontGetFontFace}
\title{cairoScaledFontGetFontFace}
\description{Gets the font face that this scaled font was created for.}
\usage{cairoScaledFontGetFontFace(scaled.font)}
\arguments{\item{\verb{scaled.font}}{[\verb{\link{CairoScaledFont}}]  a \verb{\link{CairoScaledFont}}}}
\details{  Since  1.2}
\value{[\verb{\link{CairoFontFace}}]  The \verb{\link{CairoFontFace}} with which \code{scaled.font} was
created.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
