\alias{cairoScaledFontGetFontMatrix}
\name{cairoScaledFontGetFontMatrix}
\title{cairoScaledFontGetFontMatrix}
\description{Stores the font matrix with which \code{scaled.font} was created into
\code{matrix}.}
\usage{cairoScaledFontGetFontMatrix(scaled.font)}
\arguments{\item{\verb{scaled.font}}{[\code{\link{CairoScaledFont}}]  a \code{\link{CairoScaledFont}}}}
\details{  Since 1.2}
\value{
A list containing the following elements:
\item{\verb{font.matrix}}{[\code{\link{CairoMatrix}}]  return value for the matrix}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
