\alias{cairoScaledFontGetFontOptions}
\name{cairoScaledFontGetFontOptions}
\title{cairoScaledFontGetFontOptions}
\description{Stores the font options with which \code{scaled.font} was created into
\code{options}.}
\usage{cairoScaledFontGetFontOptions(scaled.font)}
\arguments{\item{\code{scaled.font}}{[\code{\link{CairoScaledFont}}]  a \code{\link{CairoScaledFont}}}}
\details{  Since  1.2}
\value{
A list containing the following elements:
\item{\code{options}}{[\code{\link{CairoFontOptions}}]  return value for the font options}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
