\alias{cairoScaledFontGetType}
\name{cairoScaledFontGetType}
\title{cairoScaledFontGetType}
\description{This function returns the type of the backend used to create
a scaled font. See \verb{\link{CairoFontType}} for available types.}
\usage{cairoScaledFontGetType(scaled.font)}
\arguments{\item{\verb{scaled.font}}{[\verb{\link{CairoScaledFont}}]  a \verb{\link{CairoScaledFont}}}}
\details{  Since  1.2}
\value{[\verb{\link{CairoFontType}}]  The type of \code{scaled.font}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
