\alias{cairoScaledFontGlyphExtents}
\name{cairoScaledFontGlyphExtents}
\title{cairoScaledFontGlyphExtents}
\description{\emph{undocumented
}}
\usage{cairoScaledFontGlyphExtents(scaled.font, glyphs, num.glyphs)}
\arguments{
\item{\code{scaled.font}}{[\code{\link{CairoScaledFont}}] }
\item{\code{glyphs}}{[\code{\link{CairoGlyph}}] }
\item{\code{num.glyphs}}{[integer] }
}
\value{
A list containing the following elements:
\item{\code{extents}}{[\code{\link{CairoTextExtents}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
