\alias{cairoScaledFontStatus}
\name{cairoScaledFontStatus}
\title{cairoScaledFontStatus}
\description{Checks whether an error has previously occurred for this
scaled\_font.}
\usage{cairoScaledFontStatus(scaled.font)}
\arguments{\item{\code{scaled.font}}{[\code{\link{CairoScaledFont}}]  a \code{\link{CairoScaledFont}}}}
\value{[\code{\link{CairoStatus}}] \code{CAIRO_STATUS_SUCCESS} or another error such as
  \code{CAIRO_STATUS_NO_MEMORY}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
