\alias{cairoSetMatrix}
\name{cairoSetMatrix}
\title{cairoSetMatrix}
\description{Modifies the current transformation matrix (CTM) by setting it
equal to \code{matrix}.}
\usage{cairoSetMatrix(cr, matrix)}
\arguments{
\item{\code{cr}}{[\code{\link{Cairo}}]  a cairo context}
\item{\code{matrix}}{[\code{\link{CairoMatrix}}]  a transformation matrix from user space to device space}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
