\alias{cairoSetMiterLimit}
\name{cairoSetMiterLimit}
\title{cairoSetMiterLimit}
\description{\emph{undocumented
}}
\usage{cairoSetMiterLimit(cr, limit)}
\arguments{
\item{\code{cr}}{[\code{\link{Cairo}}] }
\item{\code{limit}}{[numeric] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
