\alias{cairoShowGlyphs}
\name{cairoShowGlyphs}
\title{cairoShowGlyphs}
\description{A drawing operator that generates the shape from a list of glyphs,
rendered according to the current font\_face, font\_size
(font\_matrix), and font\_options.}
\usage{cairoShowGlyphs(cr, glyphs, num.glyphs)}
\arguments{
\item{\code{cr}}{[\code{\link{Cairo}}]  a cairo context}
\item{\code{glyphs}}{[\code{\link{CairoGlyph}}]  list of glyphs to show}
\item{\code{num.glyphs}}{[integer]  number of glyphs to show}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
