\alias{cairoShowPage}
\name{cairoShowPage}
\title{cairoShowPage}
\description{Emits and clears the current page for backends that support multiple
pages.  Use \code{\link{cairoCopyPage}} if you don't want to clear the page.}
\usage{cairoShowPage(cr)}
\arguments{\item{\code{cr}}{[\code{\link{Cairo}}]  a cairo context}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
