\alias{cairoStatusToString}
\name{cairoStatusToString}
\title{cairoStatusToString}
\description{Provides a human-readable description of a \verb{\link{CairoStatus}}.}
\usage{cairoStatusToString(status)}
\arguments{\item{\verb{status}}{[\verb{\link{CairoStatus}}]  a cairo status}}
\value{[char]  a string representation of the status}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
