\alias{cairoSurfaceDestroy}
\name{cairoSurfaceDestroy}
\title{cairoSurfaceDestroy}
\description{Decreases the reference count on \code{surface} by one. If the result is
zero.  See
\code{\link{cairoSurfaceReference}}.}
\usage{cairoSurfaceDestroy(surface)}
\arguments{\item{\verb{surface}}{[\verb{\link{CairoSurface}}]  a \verb{\link{CairoSurface}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
