\alias{cairoSurfaceGetContent}
\name{cairoSurfaceGetContent}
\title{cairoSurfaceGetContent}
\description{\emph{undocumented
}}
\usage{cairoSurfaceGetContent(surface)}
\arguments{\item{\code{surface}}{[\code{\link{CairoSurface}}]  a \code{\link{CairoSurface}}}}
\details{  Since  1.2}
\value{[\code{\link{CairoContent}}]  The content type of \code{surface} which indicates whether
the surface contains color and/or alpha information. See
\code{\link{CairoContent}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
