\alias{cairoSurfaceGetType}
\name{cairoSurfaceGetType}
\title{cairoSurfaceGetType}
\description{This function returns the type of the backend used to create
a surface. See \verb{\link{CairoSurfaceType}} for available types.}
\usage{cairoSurfaceGetType(surface)}
\arguments{\item{\verb{surface}}{[\verb{\link{CairoSurface}}]  a \verb{\link{CairoSurface}}}}
\details{  Since  1.2}
\value{[\verb{\link{CairoSurfaceType}}]  The type of \code{surface}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
