\alias{cairoToyFontFaceGetFamily}
\name{cairoToyFontFaceGetFamily}
\title{cairoToyFontFaceGetFamily}
\description{Gets the familly name of a toy font.}
\usage{cairoToyFontFaceGetFamily(font.face)}
\arguments{\item{\verb{font.face}}{[\code{\link{CairoFontFace}}]  A toy font face}}
\details{  Since 1.8}
\value{[char]  The family name.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
