\alias{cairoToyFontFaceGetSlant}
\name{cairoToyFontFaceGetSlant}
\title{cairoToyFontFaceGetSlant}
\description{Gets the slant a toy font.}
\usage{cairoToyFontFaceGetSlant(font.face)}
\arguments{\item{\verb{font.face}}{[\code{\link{CairoFontFace}}]  A toy font face}}
\details{  Since 1.8}
\value{[\code{\link{CairoFontSlant}}]  The slant value}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
