\alias{cairoToyFontFaceGetWeight}
\name{cairoToyFontFaceGetWeight}
\title{cairoToyFontFaceGetWeight}
\description{Gets the weight a toy font.}
\usage{cairoToyFontFaceGetWeight(font.face)}
\arguments{\item{\verb{font.face}}{[\code{\link{CairoFontFace}}]  A toy font face}}
\details{  Since 1.8}
\value{[\code{\link{CairoFontWeight}}]  The weight value}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
