\alias{cairoVersionString}
\name{cairoVersionString}
\title{cairoVersionString}
\description{Returns the version of the cairo library as a human-readable string
of the form "X.Y.Z".}
\usage{cairoVersionString()}
\details{See also \code{\link{cairoVersion}} as well as the compile-time equivalents
\code{CAIRO_VERSION_STRING} and \code{CAIRO_VERSION}.  }
\value{[char]  a string containing the version.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
