\name{containers}
\title{GtkContainer utilities}
\alias{[[.GtkContainer}
\description{This utility gets the nth child from a \code{\link{GtkContainer}}.}
\usage{
"[[.GtkContainer(x, child)"
}
\arguments{
  \item{x}{The GtkContainer object}
  \item{child}{The index of the child to retrieve from the container}
}
\value{
  The widget at index 'child' in the vector of children in the GtkContainer
}
\author{Michael Lawrence}
\keyword{internal}
\keyword{interface}
