\alias{gdkRgbGetCmap}
\alias{gdkSelectionOwnerGet}
\alias{gdkSelectionOwnerGetForDisplay}
\alias{gdkSelectionPropertyGet}
\alias{gtkCheckMenuItemSetState}
\alias{gtkContainerChildren}
\alias{gtkLabelSet}
\alias{gtkMenuItemRightJustify}
\alias{gtkNotebookCurrentPage}
\alias{gtkNotebookSetPage}
\alias{gtkRadioButtonGroup}
\alias{gtkRadioMenuItemGroup}
\alias{gtkStyleApplyDefaultPixmap}
\alias{gtkToggleButtonSetState}
\alias{gtkTreeModelGetIterRoot}
\title{Deprecated Macros}
\name{deprecated}
\description{This is a deprecated macro. Please see the appropriate documentation for a better alternative.}
\author{Michael Lawrence}
\keyword{internal}
\keyword{interface}

