\alias{gAppInfoCanDelete}
\name{gAppInfoCanDelete}
\title{gAppInfoCanDelete}
\description{Obtains the information whether the \code{\link{GAppInfo}} can be deleted.
See \code{\link{gAppInfoDelete}}.}
\usage{gAppInfoCanDelete(object)}
\arguments{\item{\verb{object}}{a \code{\link{GAppInfo}}}}
\details{Since 2.20}
\value{[logical] \code{TRUE} if \code{appinfo} can be deleted}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
