\alias{gAppInfoCanRemoveSupportsType}
\name{gAppInfoCanRemoveSupportsType}
\title{gAppInfoCanRemoveSupportsType}
\description{Checks if a supported content type can be removed from an application.}
\usage{gAppInfoCanRemoveSupportsType(object)}
\arguments{\item{\verb{object}}{a \code{\link{GAppInfo}}.}}
\value{[logical] \code{TRUE} if it is possible to remove supported
content types from a given \code{appinfo}, \code{FALSE} if not.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
