\alias{gAppInfoDup}
\name{gAppInfoDup}
\title{gAppInfoDup}
\description{Creates a duplicate of a \code{\link{GAppInfo}}.}
\usage{gAppInfoDup(object)}
\arguments{\item{\verb{object}}{a \code{\link{GAppInfo}}.}}
\value{[\code{\link{GAppInfo}}] a duplicate of \code{appinfo}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
