\alias{gAppInfoEqual}
\name{gAppInfoEqual}
\title{gAppInfoEqual}
\description{Checks if two \code{\link{GAppInfo}}s are equal.}
\usage{gAppInfoEqual(object, appinfo2)}
\arguments{
\item{\verb{object}}{the first \code{\link{GAppInfo}}.}
\item{\verb{appinfo2}}{the second \code{\link{GAppInfo}}.}
}
\value{[logical] \code{TRUE} if \code{appinfo1} is equal to \code{appinfo2}. \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
