\alias{gAppInfoGetAllForType}
\name{gAppInfoGetAllForType}
\title{gAppInfoGetAllForType}
\description{Gets a list of all \code{\link{GAppInfo}}s for a given content type.}
\usage{gAppInfoGetAllForType(content.type)}
\arguments{\item{\verb{content.type}}{the content type to find a \code{\link{GAppInfo}} for}}
\value{[list] \verb{list} of \code{\link{GAppInfo}}s for given \code{content.type}
or \code{NULL} on error.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
