\alias{gAppInfoGetCommandline}
\name{gAppInfoGetCommandline}
\title{gAppInfoGetCommandline}
\description{Gets the commandline with which the application will be
started.}
\usage{gAppInfoGetCommandline(object)}
\arguments{\item{\verb{object}}{a \code{\link{GAppInfo}}}}
\details{Since 2.20}
\value{[char] a string containing the \code{appinfo}'s commandline,
or \code{NULL} if this information is not available}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
