\alias{gAppInfoGetDescription}
\name{gAppInfoGetDescription}
\title{gAppInfoGetDescription}
\description{Gets a human-readable description of an installed application.}
\usage{gAppInfoGetDescription(object)}
\arguments{\item{\verb{object}}{a \code{\link{GAppInfo}}.}}
\value{[char] a string containing a description of the
application \code{appinfo}, or \code{NULL} if none.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
