\alias{gAppInfoGetExecutable}
\name{gAppInfoGetExecutable}
\title{gAppInfoGetExecutable}
\description{Gets the executable's name for the installed application.}
\usage{gAppInfoGetExecutable(object)}
\arguments{\item{\verb{object}}{a \code{\link{GAppInfo}}}}
\value{[char] a string containing the \code{appinfo}'s application
binaries name}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
