\alias{gAppInfoGetIcon}
\name{gAppInfoGetIcon}
\title{gAppInfoGetIcon}
\description{Gets the icon for the application.}
\usage{gAppInfoGetIcon(object)}
\arguments{\item{\verb{object}}{a \code{\link{GAppInfo}}.}}
\value{[\code{\link{GIcon}}] the default \code{\link{GIcon}} for \code{appinfo}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
