\alias{gAppInfoGetName}
\name{gAppInfoGetName}
\title{gAppInfoGetName}
\description{Gets the installed name of the application.}
\usage{gAppInfoGetName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GAppInfo}}.}}
\value{[char] the name of the application for \code{appinfo}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
