\alias{gAppInfoResetTypeAssociations}
\name{gAppInfoResetTypeAssociations}
\title{gAppInfoResetTypeAssociations}
\description{Removes all changes to the type associations done by
\code{\link{gAppInfoSetAsDefaultForType}}, 
\code{\link{gAppInfoSetAsDefaultForExtension}}, 
\code{\link{gAppInfoAddSupportsType}} or \code{\link{gAppInfoRemoveSupportsType}}.}
\usage{gAppInfoResetTypeAssociations(content.type)}
\arguments{\item{\verb{content.type}}{a content type}}
\details{Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
