\alias{gAppInfoShouldShow}
\name{gAppInfoShouldShow}
\title{gAppInfoShouldShow}
\description{Checks if the application info should be shown in menus that 
list available applications.}
\usage{gAppInfoShouldShow(object)}
\arguments{\item{\verb{object}}{a \code{\link{GAppInfo}}.}}
\value{[logical] \code{TRUE} if the \code{appinfo} should be shown, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
