\alias{gAppInfoSupportsFiles}
\name{gAppInfoSupportsFiles}
\title{gAppInfoSupportsFiles}
\description{Checks if the application accepts files as arguments.}
\usage{gAppInfoSupportsFiles(object)}
\arguments{\item{\verb{object}}{a \code{\link{GAppInfo}}.}}
\value{[logical] \code{TRUE} if the \code{appinfo} supports files.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
