\alias{gAppInfoSupportsUris}
\name{gAppInfoSupportsUris}
\title{gAppInfoSupportsUris}
\description{Checks if the application supports reading files and directories from URIs.}
\usage{gAppInfoSupportsUris(object)}
\arguments{\item{\verb{object}}{a \code{\link{GAppInfo}}.}}
\value{[logical] \code{TRUE} if the \code{appinfo} supports URIs.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
