\alias{gAppLaunchContextNew}
\name{gAppLaunchContextNew}
\title{gAppLaunchContextNew}
\description{Creates a new application launch context. This is not normally used,
instead you instantiate a subclass of this, such as \code{\link{GdkAppLaunchContext}}.}
\usage{gAppLaunchContextNew()}
\value{[\code{\link{GAppLaunchContext}}] a \code{\link{GAppLaunchContext}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
