\alias{gAsyncResultGetSourceObject}
\name{gAsyncResultGetSourceObject}
\title{gAsyncResultGetSourceObject}
\description{Gets the source object from a \code{\link{GAsyncResult}}.}
\usage{gAsyncResultGetSourceObject(object)}
\arguments{\item{\verb{object}}{a \code{\link{GAsyncResult}}}}
\value{[\code{\link{GObject}}] a new reference to the source object for the \code{res},
or \code{NULL} if there is none.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
