\alias{gAsyncResultGetUserData}
\name{gAsyncResultGetUserData}
\title{gAsyncResultGetUserData}
\description{Gets the user data from a \code{\link{GAsyncResult}}.}
\usage{gAsyncResultGetUserData(object)}
\arguments{\item{\verb{object}}{a \code{\link{GAsyncResult}}.}}
\value{[R object] the user data for \code{res}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
