\alias{gBufferedInputStreamGetAvailable}
\name{gBufferedInputStreamGetAvailable}
\title{gBufferedInputStreamGetAvailable}
\description{Gets the size of the available data within the stream.}
\usage{gBufferedInputStreamGetAvailable(object)}
\arguments{\item{\verb{object}}{\code{\link{GBufferedInputStream}}.}}
\value{[numeric] size of the available stream.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
