\alias{gBufferedInputStreamGetBufferSize}
\name{gBufferedInputStreamGetBufferSize}
\title{gBufferedInputStreamGetBufferSize}
\description{Gets the size of the input buffer.}
\usage{gBufferedInputStreamGetBufferSize(object)}
\arguments{\item{\verb{object}}{\code{\link{GBufferedInputStream}}.}}
\value{[numeric] the current buffer size.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
