\alias{gBufferedInputStreamNew}
\name{gBufferedInputStreamNew}
\title{gBufferedInputStreamNew}
\description{Creates a new \code{\link{GInputStream}} from the given \code{base.stream}, with 
a buffer set to the default size (4 kilobytes).}
\usage{gBufferedInputStreamNew(base.stream = NULL)}
\arguments{\item{\verb{base.stream}}{a \code{\link{GInputStream}}.}}
\value{[\code{\link{GInputStream}}] a \code{\link{GInputStream}} for the given \code{base.stream}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
