\alias{gBufferedInputStreamNewSized}
\name{gBufferedInputStreamNewSized}
\title{gBufferedInputStreamNewSized}
\description{Creates a new \code{\link{GBufferedInputStream}} from the given \code{base.stream}, 
with a buffer set to \code{size}.}
\usage{gBufferedInputStreamNewSized(base.stream, size)}
\arguments{
\item{\verb{base.stream}}{a \code{\link{GInputStream}}.}
\item{\verb{size}}{a \verb{numeric}.}
}
\value{[\code{\link{GInputStream}}] a \code{\link{GInputStream}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
