\alias{gBufferedInputStreamSetBufferSize}
\name{gBufferedInputStreamSetBufferSize}
\title{gBufferedInputStreamSetBufferSize}
\description{Sets the size of the internal buffer of \code{stream} to \code{size}, or to the 
size of the contents of the buffer. The buffer can never be resized 
smaller than its current contents.}
\usage{gBufferedInputStreamSetBufferSize(object, size)}
\arguments{
\item{\verb{object}}{\code{\link{GBufferedInputStream}}.}
\item{\verb{size}}{a \verb{numeric}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
