\alias{gBufferedOutputStreamGetAutoGrow}
\name{gBufferedOutputStreamGetAutoGrow}
\title{gBufferedOutputStreamGetAutoGrow}
\description{Checks if the buffer automatically grows as data is added.}
\usage{gBufferedOutputStreamGetAutoGrow(object)}
\arguments{\item{\verb{object}}{a \code{\link{GBufferedOutputStream}}.}}
\value{[logical] \code{TRUE} if the \code{stream}'s buffer automatically grows,
\code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
