\alias{gBufferedOutputStreamGetBufferSize}
\name{gBufferedOutputStreamGetBufferSize}
\title{gBufferedOutputStreamGetBufferSize}
\description{Gets the size of the buffer in the \code{stream}.}
\usage{gBufferedOutputStreamGetBufferSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GBufferedOutputStream}}.}}
\value{[numeric] the current size of the buffer.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
