\alias{gBufferedOutputStreamNew}
\name{gBufferedOutputStreamNew}
\title{gBufferedOutputStreamNew}
\description{Creates a new buffered output stream for a base stream.}
\usage{gBufferedOutputStreamNew(base.stream = NULL)}
\arguments{\item{\verb{base.stream}}{a \code{\link{GOutputStream}}.}}
\value{[\code{\link{GOutputStream}}] a \code{\link{GOutputStream}} for the given \code{base.stream}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
