\alias{gBufferedOutputStreamNewSized}
\name{gBufferedOutputStreamNewSized}
\title{gBufferedOutputStreamNewSized}
\description{Creates a new buffered output stream with a given buffer size.}
\usage{gBufferedOutputStreamNewSized(base.stream, size)}
\arguments{
\item{\verb{base.stream}}{a \code{\link{GOutputStream}}.}
\item{\verb{size}}{a \verb{numeric}.}
}
\value{[\code{\link{GOutputStream}}] a \code{\link{GOutputStream}} with an internal buffer set to \code{size}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
