\alias{gBufferedOutputStreamSetBufferSize}
\name{gBufferedOutputStreamSetBufferSize}
\title{gBufferedOutputStreamSetBufferSize}
\description{Sets the size of the internal buffer to \code{size}.}
\usage{gBufferedOutputStreamSetBufferSize(object, size)}
\arguments{
\item{\verb{object}}{a \code{\link{GBufferedOutputStream}}.}
\item{\verb{size}}{a \verb{numeric}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
