\alias{gCancellableGetCurrent}
\name{gCancellableGetCurrent}
\title{gCancellableGetCurrent}
\description{Gets the top cancellable from the stack.}
\usage{gCancellableGetCurrent()}
\value{[\code{\link{GCancellable}}] a \code{\link{GCancellable}} from the top of the stack, or \code{NULL}
if the stack is empty.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
