\alias{gCancellableIsCancelled}
\name{gCancellableIsCancelled}
\title{gCancellableIsCancelled}
\description{Checks if a cancellable job has been cancelled.}
\usage{gCancellableIsCancelled(object)}
\arguments{\item{\verb{object}}{a \code{\link{GCancellable}} or NULL.}}
\value{[logical] \code{TRUE} if \code{cancellable} is cancelled,
FALSE if called with \code{NULL} or if item is not cancelled.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
