\alias{gCancellablePopCurrent}
\name{gCancellablePopCurrent}
\title{gCancellablePopCurrent}
\description{Pops \code{cancellable} off the cancellable stack (verifying that \code{cancellable}
is on the top of the stack).}
\usage{gCancellablePopCurrent(object)}
\arguments{\item{\verb{object}}{a \code{\link{GCancellable}} object}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
