\alias{gCancellableReset}
\name{gCancellableReset}
\title{gCancellableReset}
\description{Resets \code{cancellable} to its uncancelled state.}
\usage{gCancellableReset(object)}
\arguments{\item{\verb{object}}{a \code{\link{GCancellable}} object.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
