\alias{gContentTypeCanBeExecutable}
\name{gContentTypeCanBeExecutable}
\title{gContentTypeCanBeExecutable}
\description{Checks if a content type can be executable. Note that for instance
things like text files can be executables (i.e. scripts and batch files).}
\usage{gContentTypeCanBeExecutable(type)}
\arguments{\item{\verb{type}}{a content type string.}}
\value{[logical] \code{TRUE} if the file type corresponds to a type that
can be executable, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
