\alias{gContentTypeEquals}
\name{gContentTypeEquals}
\title{gContentTypeEquals}
\description{Compares two content types for equality.}
\usage{gContentTypeEquals(type1, type2)}
\arguments{
\item{\verb{type1}}{a content type string.}
\item{\verb{type2}}{a content type string.}
}
\value{[logical] \code{TRUE} if the two strings are identical or equivalent,
\code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
