\alias{gContentTypeFromMimeType}
\name{gContentTypeFromMimeType}
\title{gContentTypeFromMimeType}
\description{Tries to find a content type based on the mime type name.}
\usage{gContentTypeFromMimeType(mime.type)}
\arguments{\item{\verb{mime.type}}{a mime type string.}}
\details{Since 2.18}
\value{[char] Newly allocated string with content type or NULL when does not know.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
