\alias{gContentTypeGetDescription}
\name{gContentTypeGetDescription}
\title{gContentTypeGetDescription}
\description{Gets the human readable description of the content type.}
\usage{gContentTypeGetDescription(type)}
\arguments{\item{\verb{type}}{a content type string.}}
\value{[char] a short description of the content type \code{type}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
