\alias{gContentTypeGetIcon}
\name{gContentTypeGetIcon}
\title{gContentTypeGetIcon}
\description{Gets the icon for a content type.}
\usage{gContentTypeGetIcon(type)}
\arguments{\item{\verb{type}}{a content type string.}}
\value{[\code{\link{GIcon}}] \code{\link{GIcon}} corresponding to the content type.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
