\alias{gContentTypeGetMimeType}
\name{gContentTypeGetMimeType}
\title{gContentTypeGetMimeType}
\description{Gets the mime-type for the content type. If one is registered}
\usage{gContentTypeGetMimeType(type)}
\arguments{\item{\verb{type}}{a content type string.}}
\value{[char] the registered mime-type for the given \code{type}, or NULL if unknown.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
