\alias{gContentTypeIsA}
\name{gContentTypeIsA}
\title{gContentTypeIsA}
\description{Determines if \code{type} is a subset of \code{supertype}.}
\usage{gContentTypeIsA(type, supertype)}
\arguments{
\item{\verb{type}}{a content type string.}
\item{\verb{supertype}}{a string.}
}
\value{[logical] \code{TRUE} if \code{type} is a kind of \code{supertype},
\code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
