\alias{gContentTypesGetRegistered}
\name{gContentTypesGetRegistered}
\title{gContentTypesGetRegistered}
\description{Gets a list of strings containing all the registered content types
known to the system. NULL) and \code{g.list.free}(list)}
\usage{gContentTypesGetRegistered()}
\value{[list] \verb{list} of the registered content types.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
