\alias{gDataInputStreamGetByteOrder}
\name{gDataInputStreamGetByteOrder}
\title{gDataInputStreamGetByteOrder}
\description{Gets the byte order for the data input stream.}
\usage{gDataInputStreamGetByteOrder(object)}
\arguments{\item{\verb{object}}{a given \code{\link{GDataInputStream}}.}}
\value{[\code{\link{GDataStreamByteOrder}}] the \code{stream}'s current \code{\link{GDataStreamByteOrder}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
