\alias{gDataInputStreamGetNewlineType}
\name{gDataInputStreamGetNewlineType}
\title{gDataInputStreamGetNewlineType}
\description{Gets the current newline type for the \code{stream}.}
\usage{gDataInputStreamGetNewlineType(object)}
\arguments{\item{\verb{object}}{a given \code{\link{GDataInputStream}}.}}
\value{[\code{\link{GDataStreamNewlineType}}] \code{\link{GDataStreamNewlineType}} for the given \code{stream}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
