\alias{gDataInputStreamNew}
\name{gDataInputStreamNew}
\title{gDataInputStreamNew}
\description{Creates a new data input stream for the \code{base.stream}.}
\usage{gDataInputStreamNew(base.stream = NULL)}
\arguments{\item{\verb{base.stream}}{a \code{\link{GInputStream}}.}}
\value{[\code{\link{GDataInputStream}}] a new \code{\link{GDataInputStream}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
