\alias{gDataInputStreamSetByteOrder}
\name{gDataInputStreamSetByteOrder}
\title{gDataInputStreamSetByteOrder}
\description{This function sets the byte order for the given \code{stream}. All subsequent
reads from the \code{stream} will be read in the given \code{order}.}
\usage{gDataInputStreamSetByteOrder(object, order)}
\arguments{
\item{\verb{object}}{a given \code{\link{GDataInputStream}}.}
\item{\verb{order}}{a \code{\link{GDataStreamByteOrder}} to set.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
