\alias{gDataOutputStreamGetByteOrder}
\name{gDataOutputStreamGetByteOrder}
\title{gDataOutputStreamGetByteOrder}
\description{Gets the byte order for the stream.}
\usage{gDataOutputStreamGetByteOrder(object)}
\arguments{\item{\verb{object}}{a \code{\link{GDataOutputStream}}.}}
\value{[\code{\link{GDataStreamByteOrder}}] the \code{\link{GDataStreamByteOrder}} for the \code{stream}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
