\alias{gDataOutputStreamNew}
\name{gDataOutputStreamNew}
\title{gDataOutputStreamNew}
\description{Creates a new data output stream for \code{base.stream}.}
\usage{gDataOutputStreamNew(base.stream = NULL)}
\arguments{\item{\verb{base.stream}}{a \code{\link{GOutputStream}}.}}
\value{[\code{\link{GDataOutputStream}}] \code{\link{GDataOutputStream}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
