\alias{gDataOutputStreamSetByteOrder}
\name{gDataOutputStreamSetByteOrder}
\title{gDataOutputStreamSetByteOrder}
\description{Sets the byte order of the data output stream to \code{order}.}
\usage{gDataOutputStreamSetByteOrder(object, order)}
\arguments{
\item{\verb{object}}{a \code{\link{GDataOutputStream}}.}
\item{\verb{order}}{a \code{\link{GDataStreamByteOrder}}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
