\alias{gdk-pixbuf-Versioning}
\name{gdk-pixbuf-Versioning}
\title{Initialization and Versions}
\description{Library version numbers.}
\details{These functions and variables let you check the version of \command{gdk-pixbuf}
you're linking against.}
\references{\url{http://developer.gnome.org/doc/API/2.0/gdk-pixbuf/gdk-pixbuf-Versioning.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
