\alias{GDK-Pixbuf}
\name{GDK-Pixbuf}
\title{GDK-Pixbuf}
\description{This is a small library which allows you to create GdkPixbuf ('pixel buffer') objects from image data or image files. Use a \code{\link{GdkPixbuf}} in combination with \code{\link{GtkImage}} to display images.}
\details{
The RGtk binding to the GDK-Pixbuf library consists of the following components:
\describe{
\item{\link{gdk-pixbuf-animation}}{Animated images.}
\item{\link{gdk-pixbuf-creating}}{Creating a pixbuf from image data that is already in memory.}
\item{\link{gdk-pixbuf-File-Loading}}{Loading a pixbuf from a file.}
\item{\link{gdk-pixbuf-File-saving}}{Saving a pixbuf to a file.}
\item{\link{GdkPixbufLoader}}{Application-driven progressive image loading.}
\item{\link{gdk-pixbuf-gdk-pixbuf}}{Information that describes an image.}
\item{\link{gdk-pixbuf-Versioning}}{Library version numbers.}
\item{\link{gdk-pixbuf-Module-Interface}}{Extending }
\item{\link{gdk-pixbuf-scaling}}{Scaling pixbufs and scaling and compositing pixbufs}
\item{\link{gdk-pixbuf-util}}{Utility and miscellaneous convenience functions.}
}
}
\references{\url{https://developer-old.gnome.org/gdk-pixbuf/}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{interface}
