\alias{gdk-Testing}
\name{gdk-Testing}
\title{Testing}
\description{Test utilities}
\section{Methods and Functions}{
\code{\link{gdkTestRenderSync}(window)}\cr
\code{\link{gdkTestSimulateButton}(window, x, y, button, modifiers, button.pressrelease)}\cr
\code{\link{gdkTestSimulateKey}(window, x, y, keyval, modifiers, key.pressrelease)}\cr
}
\section{Detailed Description}{The functions in this section are intended to be used in test programs.
They allow to simulate some user input.}
\references{\url{http://library.gnome.org/devel//gdk/gdk-Testing.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
