\alias{gdkAppLaunchContextNew}
\name{gdkAppLaunchContextNew}
\title{gdkAppLaunchContextNew}
\description{Creates a new \code{\link{GdkAppLaunchContext}}.}
\usage{gdkAppLaunchContextNew()}
\details{Since 2.14}
\value{[\code{\link{GdkAppLaunchContext}}] a new \code{\link{GdkAppLaunchContext}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
