\alias{gdkAppLaunchContextSetDisplay}
\name{gdkAppLaunchContextSetDisplay}
\title{gdkAppLaunchContextSetDisplay}
\description{Sets the display on which applications will be launched when
using this context. See also \code{\link{gdkAppLaunchContextSetScreen}}.}
\usage{gdkAppLaunchContextSetDisplay(object, display)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkAppLaunchContext}}}
\item{\verb{display}}{a \code{\link{GdkDisplay}}}
}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
