\name{gdkAtoms}
\title{Gdk Atom Constants}
\alias{GDK_SELECTION_PRIMARY}
\alias{GDK_SELECTION_SECONDARY}
\alias{GDK_SELECTION_CLIPBOARD}
\alias{GDK_TARGET_BITMAP}
\alias{GDK_TARGET_COLORMAP}
\alias{GDK_TARGET_DRAWABLE}
\alias{GDK_TARGET_PIXMAP}
\alias{GDK_TARGET_STRING}
\alias{GDK_SELECTION_TYPE_ATOM}
\alias{GDK_SELECTION_TYPE_BITMAP}
\alias{GDK_SELECTION_TYPE_COLORMAP}
\alias{GDK_SELECTION_TYPE_DRAWABLE}
\alias{GDK_SELECTION_TYPE_INTEGER}
\alias{GDK_SELECTION_TYPE_PIXMAP}
\alias{GDK_SELECTION_TYPE_WINDOW}
\alias{GDK_SELECTION_TYPE_STRING}
\description{This is a numeric value that represents a
  \code{\link{GdkAtom}}, which is a type of identifier used by certain
  low-level systems, such as the clipboard.}
\author{Michael Lawrence}
\keyword{internal}