\alias{gdkCairoRegion}
\name{gdkCairoRegion}
\title{gdkCairoRegion}
\description{Adds the given region to the current path of \code{cr}.}
\usage{gdkCairoRegion(cr, region)}
\arguments{
\item{\code{cr}}{[\code{\link{cairo_t}}]  a \code{\link{cairo_t}}}
\item{\code{region}}{[\code{\link{GdkRegion}}]  a \code{\link{GdkRegion}}}
}
\details{  Since  2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
