\alias{gdkCairoSetSourceColor}
\name{gdkCairoSetSourceColor}
\title{gdkCairoSetSourceColor}
\description{Sets the specified \verb{\link{GdkColor}} as the source color of \code{cr}.}
\usage{gdkCairoSetSourceColor(cr, color)}
\arguments{
\item{\verb{cr}}{[\verb{\link{Cairo}}]  a \verb{\link{Cairo}}}
\item{\verb{color}}{[\verb{\link{GdkColor}}]  a \verb{\link{GdkColor}}}
}
\details{  Since  2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
