\alias{gdkColormapFreeColors}
\name{gdkColormapFreeColors}
\title{gdkColormapFreeColors}
\description{Frees previously allocated colors.}
\usage{gdkColormapFreeColors(object, colors)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkColormap}}]  a \verb{\link{GdkColormap}}.}
\item{\verb{colors}}{[\verb{\link{GdkColor}}]  the colors to free.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
