\alias{gdkColormapGetScreen}
\name{gdkColormapGetScreen}
\title{gdkColormapGetScreen}
\description{Gets the screen for which this colormap was created.}
\usage{gdkColormapGetScreen(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkColormap}}]  a \verb{\link{GdkColormap}}}}
\details{  Since  2.2}
\value{[\verb{\link{GdkScreen}}]  the screen for which this colormap was created.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
