\alias{gdkColormapGetSystemSize}
\name{gdkColormapGetSystemSize}
\title{gdkColormapGetSystemSize}
\description{
Returns the size of the system's default colormap.
(See the description of struct \code{\link{GdkColormap}} for an
explanation of the size of a colormap.)
\strong{WARNING: \code{gdk_colormap_get_system_size} is deprecated and should not be used in newly-written code.}
}
\usage{gdkColormapGetSystemSize()}
\value{[integer] the size of the system's default colormap.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
