\alias{gdkColorsStore}
\name{gdkColorsStore}
\title{gdkColorsStore}
\description{
Changes the value of the first \code{ncolors} colors in
a private colormap. This function is obsolete and
should not be used. See \code{\link{gdkColorChange}}.
\strong{WARNING: \code{gdk_colors_store} is deprecated and should not be used in newly-written code.}
}
\usage{gdkColorsStore(object, colors)}
\arguments{\item{\verb{colors}}{the new color values.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
