\alias{gdkColorToString}
\name{gdkColorToString}
\title{gdkColorToString}
\description{Returns a textual specification of \code{color} in the hexadecimal form
\code{#rrrrggggbbbb}, where \code{r},
\code{g} and \code{b} are hex digits
representing the red, green and blue components respectively.}
\usage{gdkColorToString(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkColor}}}}
\details{Since 2.12}
\value{[character] a newly-allocated text string}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
