\alias{gdkCursorGetDisplay}
\name{gdkCursorGetDisplay}
\title{gdkCursorGetDisplay}
\description{Returns the display on which the \verb{\link{GdkCursor}} is defined.}
\usage{gdkCursorGetDisplay(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkCursor}}]  a \verb{\link{GdkCursor}}.}}
\details{  Since  2.2}
\value{[\verb{\link{GdkDisplay}}]  the \verb{\link{GdkDisplay}} associated to \code{cursor}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
