\alias{gdkDeviceGetCorePointer}
\name{gdkDeviceGetCorePointer}
\title{gdkDeviceGetCorePointer}
\description{Returns the core pointer device for the default display.}
\usage{gdkDeviceGetCorePointer()}
\value{[\code{\link{GdkDevice}}] .}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
