\alias{gdkDeviceSetAxisUse}
\name{gdkDeviceSetAxisUse}
\title{gdkDeviceSetAxisUse}
\description{Specifies how an axis of a device is used.}
\usage{gdkDeviceSetAxisUse(object, index, use)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkDevice}}] a \verb{\link{GdkDevice}}.}
\item{\verb{index}}{[numeric] the index of the axis.}
\item{\verb{use}}{[\code{\link{GdkAxisUse}}] specifies how the axis is used.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
