\alias{gdkDeviceSetSource}
\name{gdkDeviceSetSource}
\title{gdkDeviceSetSource}
\description{Sets the source type for an input device.}
\usage{gdkDeviceSetSource(object, source)}
\arguments{
\item{\code{object}}{[\code{\link{GdkDevice}}] a \code{\link{GdkDevice}}.}
\item{\code{source}}{[\code{\link{GdkInputSource}}] the source type.}
}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
