\alias{gdkDevicesList}
\name{gdkDevicesList}
\title{gdkDevicesList}
\description{Returns the list of available input devices for the default display.}
\usage{gdkDevicesList()}
\value{[list]  a list of \code{\link{GdkDevice}}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
