\alias{gdkDisplayBeep}
\name{gdkDisplayBeep}
\title{gdkDisplayBeep}
\description{Emits a short beep on \code{display}}
\usage{gdkDisplayBeep(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkDisplay}}]  a \verb{\link{GdkDisplay}}}}
\details{  Since  2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
