\alias{gdkDisplayGetDefaultCursorSize}
\name{gdkDisplayGetDefaultCursorSize}
\title{gdkDisplayGetDefaultCursorSize}
\description{Returns the default size to use for cursors on \code{display}.}
\usage{gdkDisplayGetDefaultCursorSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkDisplay}}}}
\details{Since 2.4}
\value{[numeric] the default cursor size.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
