\alias{gdkDisplayGetDefaultScreen}
\name{gdkDisplayGetDefaultScreen}
\title{gdkDisplayGetDefaultScreen}
\description{Get the default \verb{\link{GdkScreen}} for \code{display}.}
\usage{gdkDisplayGetDefaultScreen(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkDisplay}}]  a \verb{\link{GdkDisplay}}}}
\details{  Since  2.2}
\value{[\verb{\link{GdkScreen}}]  the default \verb{\link{GdkScreen}} object for \code{display}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
