\alias{gdkDisplayGetName}
\name{gdkDisplayGetName}
\title{gdkDisplayGetName}
\description{Gets the name of the display.}
\usage{gdkDisplayGetName(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkDisplay}}]  a \verb{\link{GdkDisplay}}}}
\details{  Since  2.2}
\value{[character]  a string representing the display name.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
