\alias{gdkDisplayGetNScreens}
\name{gdkDisplayGetNScreens}
\title{gdkDisplayGetNScreens}
\description{Gets the number of screen managed by the \code{display}.}
\usage{gdkDisplayGetNScreens(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkDisplay}}}}
\details{Since 2.2}
\value{[integer] number of screens.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
