\alias{gdkDisplayGetScreen}
\name{gdkDisplayGetScreen}
\title{gdkDisplayGetScreen}
\description{Returns a screen object for one of the screens of the display.}
\usage{gdkDisplayGetScreen(object, screen.num)}
\arguments{
\item{\code{object}}{[\code{\link{GdkDisplay}}]  a \code{\link{GdkDisplay}}}
\item{\code{screen.num}}{[integer]  the screen number}
}
\details{  Since  2.2}
\value{[\code{\link{GdkScreen}}]  the \code{\link{GdkScreen}} object}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
