\alias{gdkDisplayKeyboardUngrab}
\name{gdkDisplayKeyboardUngrab}
\title{gdkDisplayKeyboardUngrab}
\description{Release any keyboard grab}
\usage{gdkDisplayKeyboardUngrab(object, time. = GDK_CURRENT_TIME)}
\arguments{
\item{\code{object}}{[\code{\link{GdkDisplay}}]  a \code{\link{GdkDisplay}}.}
\item{\code{time.}}{[numeric]  a timestap (e.g \code{GDK_CURRENT_TIME}).}
}
\details{  Since  2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
