\alias{gdkDisplayManagerSetDefaultDisplay}
\name{gdkDisplayManagerSetDefaultDisplay}
\title{gdkDisplayManagerSetDefaultDisplay}
\description{Sets \code{display} as the default display.}
\usage{gdkDisplayManagerSetDefaultDisplay(object, display)}
\arguments{
\item{\code{object}}{[\code{\link{GdkDisplayManager}}]  a \code{\link{GdkDisplayManager}}}
\item{\code{display}}{[\code{\link{GdkDisplay}}]  a \code{\link{GdkDisplay}}}
}
\details{  Since  2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
