\alias{gdkDisplayOpen}
\name{gdkDisplayOpen}
\title{gdkDisplayOpen}
\description{Opens a display.}
\usage{gdkDisplayOpen(display.name)}
\arguments{\item{\verb{display.name}}{[character]  the name of the display to open}}
\details{  Since  2.2}
\value{[\verb{\link{GdkDisplay}}]  a \verb{\link{GdkDisplay}}, or \code{NULL} if the display
 could not be opened.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
