\alias{gdkDisplayPointerIsGrabbed}
\name{gdkDisplayPointerIsGrabbed}
\title{gdkDisplayPointerIsGrabbed}
\description{Test if the pointer is grabbed.}
\usage{gdkDisplayPointerIsGrabbed(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkDisplay}}}}
\details{Since 2.2}
\value{[logical] \code{TRUE} if an active X pointer grab is in effect}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
