\alias{gdkDisplayPointerUngrab}
\name{gdkDisplayPointerUngrab}
\title{gdkDisplayPointerUngrab}
\description{Release any pointer grab.}
\usage{gdkDisplayPointerUngrab(object, time. = GDK_CURRENT_TIME)}
\arguments{
\item{\code{object}}{[\code{\link{GdkDisplay}}]  a \code{\link{GdkDisplay}}.}
\item{\code{time.}}{[numeric]  a timestap (e.g. GDK_CURRENT_TIME).}
}
\details{  Since  2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
