\alias{gdkDisplaySupportsCursorColor}
\name{gdkDisplaySupportsCursorColor}
\title{gdkDisplaySupportsCursorColor}
\description{Returns \code{TRUE} if multicolored cursors are supported
on \code{display}. Otherwise, cursors have only a forground
and a background color.}
\usage{gdkDisplaySupportsCursorColor(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkDisplay}}}}
\details{Since 2.4}
\value{[logical] whether cursors can have multiple colors.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
