\alias{gdkDisplaySupportsInputShapes}
\name{gdkDisplaySupportsInputShapes}
\title{gdkDisplaySupportsInputShapes}
\description{Returns \code{TRUE} if \code{\link{gdkWindowInputShapeCombineMask}} can
be used to modify the input shape of windows on \code{display}.}
\usage{gdkDisplaySupportsInputShapes(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkDisplay}}}}
\details{Since 2.10}
\value{[logical] \code{TRUE} if windows with modified input shape are supported}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
