\alias{gdkDisplaySupportsShapes}
\name{gdkDisplaySupportsShapes}
\title{gdkDisplaySupportsShapes}
\description{Returns \code{TRUE} if \code{\link{gdkWindowShapeCombineMask}} can
be used to create shaped windows on \code{display}.}
\usage{gdkDisplaySupportsShapes(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkDisplay}}}}
\details{Since 2.10}
\value{[logical] \code{TRUE} if shaped windows are supported}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
