\alias{gdkDragContextNew}
\name{gdkDragContextNew}
\title{gdkDragContextNew}
\description{Creates a new \verb{\link{GdkDragContext}}.}
\usage{gdkDragContextNew()}
\value{[\verb{\link{GdkDragContext}}]  the newly created \verb{\link{GdkDragContext}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
