\alias{gdkDragGetSelection}
\name{gdkDragGetSelection}
\title{gdkDragGetSelection}
\description{Returns the selection atom for the current source window.}
\usage{gdkDragGetSelection(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkDragContext}}]  a \verb{\link{GdkDragContext}}.}}
\value{[\verb{\link{GdkAtom}}]  the selection atom.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
