\alias{gdkDrawableGetColormap}
\name{gdkDrawableGetColormap}
\title{gdkDrawableGetColormap}
\description{Gets the colormap for \code{drawable}, if one is set; returns
\code{NULL} otherwise.}
\usage{gdkDrawableGetColormap(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkDrawable}}]  a \verb{\link{GdkDrawable}}}}
\value{[\verb{\link{GdkColormap}}]  the colormap, or \code{NULL}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
