\alias{gdkDrawableGetDepth}
\name{gdkDrawableGetDepth}
\title{gdkDrawableGetDepth}
\description{Obtains the bit depth of the drawable, that is, the number of bits
that make up a pixel in the drawable's visual. Examples are 8 bits
per pixel, 24 bits per pixel, etc.}
\usage{gdkDrawableGetDepth(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkDrawable}}}}
\value{[integer] number of bits per pixel}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
