\alias{gdkDrawableGetDisplay}
\name{gdkDrawableGetDisplay}
\title{gdkDrawableGetDisplay}
\description{Gets the \verb{\link{GdkDisplay}} associated with a \verb{\link{GdkDrawable}}.}
\usage{gdkDrawableGetDisplay(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkDrawable}}]  a \verb{\link{GdkDrawable}}}}
\details{  Since  2.2}
\value{[\verb{\link{GdkDisplay}}]  the \verb{\link{GdkDisplay}} associated with \code{drawable}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
