\alias{gdkDrawableGetScreen}
\name{gdkDrawableGetScreen}
\title{gdkDrawableGetScreen}
\description{Gets the \verb{\link{GdkScreen}} associated with a \verb{\link{GdkDrawable}}.}
\usage{gdkDrawableGetScreen(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkDrawable}}]  a \verb{\link{GdkDrawable}}}}
\details{  Since  2.2}
\value{[\verb{\link{GdkScreen}}]  the \verb{\link{GdkScreen}} associated with \code{drawable}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
