\alias{gdkDrawableGetVisual}
\name{gdkDrawableGetVisual}
\title{gdkDrawableGetVisual}
\description{Gets the \code{\link{GdkVisual}} describing the pixel format of \code{drawable}.}
\usage{gdkDrawableGetVisual(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkDrawable}}]  a \code{\link{GdkDrawable}}}}
\value{[\code{\link{GdkVisual}}]  a \code{\link{GdkVisual}}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
