\alias{gdkDrawPoint}
\name{gdkDrawPoint}
\title{gdkDrawPoint}
\description{Draws a point, using the foreground color and other attributes of 
the \code{\link{GdkGC}}.}
\usage{gdkDrawPoint(object, gc, x, y)}
\arguments{
\item{\code{object}}{[\code{\link{GdkDrawable}}]  a \code{\link{GdkDrawable}} (a \code{\link{GdkWindow}} or a \code{\link{GdkPixmap}}).}
\item{\code{gc}}{[\code{\link{GdkGC}}]  a \code{\link{GdkGC}}.}
\item{\code{x}}{[integer]  the x coordinate of the point.}
\item{\code{y}}{[integer]  the y coordinate of the point.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
