\alias{gdkEventCopy}
\name{gdkEventCopy}
\title{gdkEventCopy}
\description{Copies a \verb{\link{GdkEvent}}, copying or incrementing the reference count of the
resources associated with it (e.g. \verb{\link{GdkWindow}}'s and strings).}
\usage{gdkEventCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkEvent}}]  a \verb{\link{GdkEvent}}}}
\value{[\verb{\link{GdkEvent}}]  a copy of \code{event}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
