\alias{gdkEventPut}
\name{gdkEventPut}
\title{gdkEventPut}
\description{Appends a copy of the given event onto the front of the event
queue for event->any.window's display, or the default event
queue if event->any.window is \code{NULL}. See \code{\link{gdkDisplayPutEvent}}.}
\usage{gdkEventPut(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkEvent}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
