\alias{gdkEventSetScreen}
\name{gdkEventSetScreen}
\title{gdkEventSetScreen}
\description{Sets the screen for \code{event} to \code{screen}. The event must
have been allocated by GTK+, for instance, by
\code{\link{gdkEventCopy}}.}
\usage{gdkEventSetScreen(object, screen)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkEvent}}]  a \verb{\link{GdkEvent}}}
\item{\verb{screen}}{[\verb{\link{GdkScreen}}]  a \verb{\link{GdkScreen}}}
}
\details{  Since  2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
