\alias{gdkEventsPending}
\name{gdkEventsPending}
\title{gdkEventsPending}
\description{Checks if any events are ready to be processed for any display.}
\usage{gdkEventsPending()}
\value{[logical] \code{TRUE} if any events are pending.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
