\alias{gdkFontGetDisplay}
\name{gdkFontGetDisplay}
\title{gdkFontGetDisplay}
\description{
Returns the \verb{\link{GdkDisplay}} for \code{font}.
\strong{WARNING: \code{gdk_font_get_display} is deprecated and should not be used in newly-written code.}
}
\usage{gdkFontGetDisplay(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkFont}}]  the \verb{\link{GdkFont}}.}}
\details{  Since  2.2}
\value{[\verb{\link{GdkDisplay}}]  the corresponding \verb{\link{GdkDisplay}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
