\alias{gdkFontId}
\name{gdkFontId}
\title{gdkFontId}
\description{
Returns the X Font ID for the given font.
\strong{WARNING: \code{gdk_font_id} is deprecated and should not be used in newly-written code.}
}
\usage{gdkFontId(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkFont}}.}}
\value{[integer] the numeric X Font ID}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
