\alias{gdkGCCopy}
\name{gdkGCCopy}
\title{gdkGCCopy}
\description{Copy the set of values from one graphics context
onto another graphics context.}
\usage{gdkGCCopy(object, src.gc)}
\arguments{
\item{\verb{object}}{the destination graphics context.}
\item{\verb{src.gc}}{the source graphics context.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
