\alias{gdkGCGetScreen}
\name{gdkGCGetScreen}
\title{gdkGCGetScreen}
\description{Gets the \verb{\link{GdkScreen}} for which \code{gc} was created}
\usage{gdkGCGetScreen(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkGC}}]  a \verb{\link{GdkGC}}.}}
\details{  Since  2.2}
\value{[\verb{\link{GdkScreen}}]  the \verb{\link{GdkScreen}} for \code{gc}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
