\alias{gdkGCSetClipMask}
\name{gdkGCSetClipMask}
\title{gdkGCSetClipMask}
\description{Sets the clip mask for a graphics context from a bitmap.
The clip mask is interpreted relative to the clip
origin. (See \code{\link{gdkGCSetClipOrigin}}).}
\usage{gdkGCSetClipMask(object, mask)}
\arguments{
\item{\verb{object}}{the \code{\link{GdkGC}}.}
\item{\verb{mask}}{a bitmap.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
