\alias{gdkGCSetClipRectangle}
\name{gdkGCSetClipRectangle}
\title{gdkGCSetClipRectangle}
\description{Sets the clip mask for a graphics context from a
rectangle. The clip mask is interpreted relative to the clip
origin. (See \code{\link{gdkGCSetClipOrigin}}).}
\usage{gdkGCSetClipRectangle(object, rectangle)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkGC}}]  a \verb{\link{GdkGC}}.}
\item{\verb{rectangle}}{[\verb{\link{GdkRectangle}}]  the rectangle to clip to.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
