\alias{gdkGCSetClipRegion}
\name{gdkGCSetClipRegion}
\title{gdkGCSetClipRegion}
\description{Sets the clip mask for a graphics context from a region structure.
The clip mask is interpreted relative to the clip origin. (See
\code{\link{gdkGCSetClipOrigin}}).}
\usage{gdkGCSetClipRegion(object, region)}
\arguments{
\item{\code{object}}{[\code{\link{GdkGC}}]  a \code{\link{GdkGC}}.}
\item{\code{region}}{[\code{\link{GdkRegion}}]  the \code{\link{GdkRegion}}. }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
