\alias{gdkGCSetFill}
\name{gdkGCSetFill}
\title{gdkGCSetFill}
\description{Set the fill mode for a graphics context.}
\usage{gdkGCSetFill(object, fill)}
\arguments{
\item{\code{object}}{[\code{\link{GdkGC}}]  a \code{\link{GdkGC}}.}
\item{\code{fill}}{[\code{\link{GdkFill}}]  the new fill mode.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
