\alias{gdkGCSetFunction}
\name{gdkGCSetFunction}
\title{gdkGCSetFunction}
\description{Determines how the current pixel values and the
pixel values being drawn are combined to produce
the final pixel values.}
\usage{gdkGCSetFunction(object, fun)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkGC}}]  a \verb{\link{GdkGC}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
