\alias{gdkGCSetStipple}
\name{gdkGCSetStipple}
\title{gdkGCSetStipple}
\description{Set the stipple bitmap for a graphics context. The
stipple will only be used if the fill mode is
\code{GDK_STIPPLED} or \code{GDK_OPAQUE_STIPPLED}.}
\usage{gdkGCSetStipple(object, stipple)}
\arguments{
\item{\code{object}}{[\code{\link{GdkGC}}]  a \code{\link{GdkGC}}.}
\item{\code{stipple}}{[\code{\link{GdkPixmap}}]  the new stipple bitmap.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
