\alias{gdkGetDefaultRootWindow}
\name{gdkGetDefaultRootWindow}
\title{gdkGetDefaultRootWindow}
\description{Obtains the root window (parent all other windows are inside)
for the default display and screen.}
\usage{gdkGetDefaultRootWindow()}
\value{[\code{\link{GdkWindow}}] the default root window}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
