\alias{gdkGetDisplay}
\name{gdkGetDisplay}
\title{gdkGetDisplay}
\description{Gets the name of the display, which usually comes from the \env{DISPLAY}
environment variable or the \option{--display} command line option.}
\usage{gdkGetDisplay()}
\value{[character] the name of the display.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
