\alias{gdkGetDisplayArgName}
\name{gdkGetDisplayArgName}
\title{gdkGetDisplayArgName}
\description{Gets the display name specified in the command line arguments passed
to \code{gdkInit()} or \code{gdkParseArgs()}, if any.}
\usage{gdkGetDisplayArgName()}
\details{  Since  2.2}
\value{[character] .}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
