\alias{gdkGetProgramClass}
\name{gdkGetProgramClass}
\title{gdkGetProgramClass}
\description{Gets the program class. Unless the program class has explicitly
been set with \code{\link{gdkSetProgramClass}} or with the \option{--class}
commandline option, the default value is the program name (determined
with \code{gGetPrgname()}) with the first character converted to uppercase.}
\usage{gdkGetProgramClass()}
\value{[char] the program class.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
