\alias{gdkGetShowEvents}
\name{gdkGetShowEvents}
\title{gdkGetShowEvents}
\description{Gets whether event debugging output is enabled.}
\usage{gdkGetShowEvents()}
\value{[logical] \code{TRUE} if event debugging output is enabled.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
