\alias{gdkImageGetPixel}
\name{gdkImageGetPixel}
\title{gdkImageGetPixel}
\description{Gets a pixel value at a specified position in a \verb{\link{GdkImage}}.}
\usage{gdkImageGetPixel(object, x, y)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkImage}}] a \verb{\link{GdkImage}}.}
\item{\verb{x}}{[integer] the x coordinate of the pixel to get.}
\item{\verb{y}}{[integer] the y coordinate of the pixel to get.}
}
\value{[numeric] the pixel value at the given position.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
