\alias{gdkImagePutPixel}
\name{gdkImagePutPixel}
\title{gdkImagePutPixel}
\description{Sets a pixel in a \code{\link{GdkImage}} to a given pixel value.}
\usage{gdkImagePutPixel(object, x, y, pixel)}
\arguments{
\item{\code{object}}{[\code{\link{GdkImage}}] a \code{\link{GdkImage}}.}
\item{\code{x}}{[integer] the x coordinate of the pixel to set.}
\item{\code{y}}{[integer] the y coordinate of the pixel to set.}
\item{\code{pixel}}{[numeric] the pixel value to set.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
