\alias{gdkInputSetExtensionEvents}
\name{gdkInputSetExtensionEvents}
\title{gdkInputSetExtensionEvents}
\description{Turns extension events on or off for a particular window,
and specifies the event mask for extension events.}
\usage{gdkInputSetExtensionEvents(object, mask, mode)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkWindow}}] a \verb{\link{GdkWindow}}.}
\item{\verb{mask}}{[integer] the event mask}
\item{\verb{mode}}{[\code{\link{GdkExtensionMode}}] the type of extension events that are desired.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
