\alias{gdkKeymapGetCapsLockState}
\name{gdkKeymapGetCapsLockState}
\title{gdkKeymapGetCapsLockState}
\description{Returns whether the Caps Lock modifer is locked.}
\usage{gdkKeymapGetCapsLockState(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkKeymap}}}}
\details{Since 2.16}
\value{[logical] \code{TRUE} if Caps Lock is on}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
