\alias{gdkKeymapGetDefault}
\name{gdkKeymapGetDefault}
\title{gdkKeymapGetDefault}
\description{Returns the \verb{\link{GdkKeymap}} attached to the default display.}
\usage{gdkKeymapGetDefault()}
\value{[\verb{\link{GdkKeymap}}]  the \verb{\link{GdkKeymap}} attached to the default display.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
