\alias{gdkKeymapGetForDisplay}
\name{gdkKeymapGetForDisplay}
\title{gdkKeymapGetForDisplay}
\description{Returns the \code{\link{GdkKeymap}} attached to \code{display}.}
\usage{gdkKeymapGetForDisplay(display)}
\arguments{\item{\code{display}}{[\code{\link{GdkDisplay}}]  the \code{\link{GdkDisplay}}.}}
\details{  Since  2.2}
\value{[\code{\link{GdkKeymap}}]  the \code{\link{GdkKeymap}} attached to \code{display}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
