\alias{gdkKeymapHaveBidiLayouts}
\name{gdkKeymapHaveBidiLayouts}
\title{gdkKeymapHaveBidiLayouts}
\description{Determines if keyboard layouts for both right-to-left and left-to-right
languages are in use.}
\usage{gdkKeymapHaveBidiLayouts(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkKeymap}}]  a \code{\link{GdkKeymap}} or \code{NULL} to use the default keymap}}
\details{\code{Returns}: \code{TRUE} if there are layouts in both directions, \code{FALSE} otherwise  

  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
