\alias{GDK_VoidSymbol}
\alias{GDK_BackSpace}
\alias{GDK_Tab}
\alias{GDK_Linefeed}
\alias{GDK_Clear}
\alias{GDK_Return}
\alias{GDK_Pause}
\alias{GDK_Scroll_Lock}
\alias{GDK_Sys_Req}
\alias{GDK_Escape}
\alias{GDK_Delete}
\alias{GDK_Multi_key}
\alias{GDK_Codeinput}
\alias{GDK_SingleCandidate}
\alias{GDK_MultipleCandidate}
\alias{GDK_PreviousCandidate}
\alias{GDK_Kanji}
\alias{GDK_Muhenkan}
\alias{GDK_Henkan_Mode}
\alias{GDK_Henkan}
\alias{GDK_Romaji}
\alias{GDK_Hiragana}
\alias{GDK_Katakana}
\alias{GDK_Hiragana_Katakana}
\alias{GDK_Zenkaku}
\alias{GDK_Hankaku}
\alias{GDK_Zenkaku_Hankaku}
\alias{GDK_Touroku}
\alias{GDK_Massyo}
\alias{GDK_Kana_Lock}
\alias{GDK_Kana_Shift}
\alias{GDK_Eisu_Shift}
\alias{GDK_Eisu_toggle}
\alias{GDK_Kanji_Bangou}
\alias{GDK_Zen_Koho}
\alias{GDK_Mae_Koho}
\alias{GDK_Home}
\alias{GDK_Left}
\alias{GDK_Up}
\alias{GDK_Right}
\alias{GDK_Down}
\alias{GDK_Prior}
\alias{GDK_Page_Up}
\alias{GDK_Next}
\alias{GDK_Page_Down}
\alias{GDK_End}
\alias{GDK_Begin}
\alias{GDK_Select}
\alias{GDK_Print}
\alias{GDK_Execute}
\alias{GDK_Insert}
\alias{GDK_Undo}
\alias{GDK_Redo}
\alias{GDK_Menu}
\alias{GDK_Find}
\alias{GDK_Cancel}
\alias{GDK_Help}
\alias{GDK_Break}
\alias{GDK_Mode_switch}
\alias{GDK_script_switch}
\alias{GDK_Num_Lock}
\alias{GDK_KP_Space}
\alias{GDK_KP_Tab}
\alias{GDK_KP_Enter}
\alias{GDK_KP_F1}
\alias{GDK_KP_F2}
\alias{GDK_KP_F3}
\alias{GDK_KP_F4}
\alias{GDK_KP_Home}
\alias{GDK_KP_Left}
\alias{GDK_KP_Up}
\alias{GDK_KP_Right}
\alias{GDK_KP_Down}
\alias{GDK_KP_Prior}
\alias{GDK_KP_Page_Up}
\alias{GDK_KP_Next}
\alias{GDK_KP_Page_Down}
\alias{GDK_KP_End}
\alias{GDK_KP_Begin}
\alias{GDK_KP_Insert}
\alias{GDK_KP_Delete}
\alias{GDK_KP_Equal}
\alias{GDK_KP_Multiply}
\alias{GDK_KP_Add}
\alias{GDK_KP_Separator}
\alias{GDK_KP_Subtract}
\alias{GDK_KP_Decimal}
\alias{GDK_KP_Divide}
\alias{GDK_KP_0}
\alias{GDK_KP_1}
\alias{GDK_KP_2}
\alias{GDK_KP_3}
\alias{GDK_KP_4}
\alias{GDK_KP_5}
\alias{GDK_KP_6}
\alias{GDK_KP_7}
\alias{GDK_KP_8}
\alias{GDK_KP_9}
\alias{GDK_F1}
\alias{GDK_F2}
\alias{GDK_F3}
\alias{GDK_F4}
\alias{GDK_F5}
\alias{GDK_F6}
\alias{GDK_F7}
\alias{GDK_F8}
\alias{GDK_F9}
\alias{GDK_F10}
\alias{GDK_F11}
\alias{GDK_L1}
\alias{GDK_F12}
\alias{GDK_L2}
\alias{GDK_F13}
\alias{GDK_L3}
\alias{GDK_F14}
\alias{GDK_L4}
\alias{GDK_F15}
\alias{GDK_L5}
\alias{GDK_F16}
\alias{GDK_L6}
\alias{GDK_F17}
\alias{GDK_L7}
\alias{GDK_F18}
\alias{GDK_L8}
\alias{GDK_F19}
\alias{GDK_L9}
\alias{GDK_F20}
\alias{GDK_L10}
\alias{GDK_F21}
\alias{GDK_R1}
\alias{GDK_F22}
\alias{GDK_R2}
\alias{GDK_F23}
\alias{GDK_R3}
\alias{GDK_F24}
\alias{GDK_R4}
\alias{GDK_F25}
\alias{GDK_R5}
\alias{GDK_F26}
\alias{GDK_R6}
\alias{GDK_F27}
\alias{GDK_R7}
\alias{GDK_F28}
\alias{GDK_R8}
\alias{GDK_F29}
\alias{GDK_R9}
\alias{GDK_F30}
\alias{GDK_R10}
\alias{GDK_F31}
\alias{GDK_R11}
\alias{GDK_F32}
\alias{GDK_R12}
\alias{GDK_F33}
\alias{GDK_R13}
\alias{GDK_F34}
\alias{GDK_R14}
\alias{GDK_F35}
\alias{GDK_R15}
\alias{GDK_Shift_L}
\alias{GDK_Shift_R}
\alias{GDK_Control_L}
\alias{GDK_Control_R}
\alias{GDK_Caps_Lock}
\alias{GDK_Shift_Lock}
\alias{GDK_Meta_L}
\alias{GDK_Meta_R}
\alias{GDK_Alt_L}
\alias{GDK_Alt_R}
\alias{GDK_Super_L}
\alias{GDK_Super_R}
\alias{GDK_Hyper_L}
\alias{GDK_Hyper_R}
\alias{GDK_ISO_Lock}
\alias{GDK_ISO_Level2_Latch}
\alias{GDK_ISO_Level3_Shift}
\alias{GDK_ISO_Level3_Latch}
\alias{GDK_ISO_Level3_Lock}
\alias{GDK_ISO_Group_Shift}
\alias{GDK_ISO_Group_Latch}
\alias{GDK_ISO_Group_Lock}
\alias{GDK_ISO_Next_Group}
\alias{GDK_ISO_Next_Group_Lock}
\alias{GDK_ISO_Prev_Group}
\alias{GDK_ISO_Prev_Group_Lock}
\alias{GDK_ISO_First_Group}
\alias{GDK_ISO_First_Group_Lock}
\alias{GDK_ISO_Last_Group}
\alias{GDK_ISO_Last_Group_Lock}
\alias{GDK_ISO_Left_Tab}
\alias{GDK_ISO_Move_Line_Up}
\alias{GDK_ISO_Move_Line_Down}
\alias{GDK_ISO_Partial_Line_Up}
\alias{GDK_ISO_Partial_Line_Down}
\alias{GDK_ISO_Partial_Space_Left}
\alias{GDK_ISO_Partial_Space_Right}
\alias{GDK_ISO_Set_Margin_Left}
\alias{GDK_ISO_Set_Margin_Right}
\alias{GDK_ISO_Release_Margin_Left}
\alias{GDK_ISO_Release_Margin_Right}
\alias{GDK_ISO_Release_Both_Margins}
\alias{GDK_ISO_Fast_Cursor_Left}
\alias{GDK_ISO_Fast_Cursor_Right}
\alias{GDK_ISO_Fast_Cursor_Up}
\alias{GDK_ISO_Fast_Cursor_Down}
\alias{GDK_ISO_Continuous_Underline}
\alias{GDK_ISO_Discontinuous_Underline}
\alias{GDK_ISO_Emphasize}
\alias{GDK_ISO_Center_Object}
\alias{GDK_ISO_Enter}
\alias{GDK_dead_grave}
\alias{GDK_dead_acute}
\alias{GDK_dead_circumflex}
\alias{GDK_dead_tilde}
\alias{GDK_dead_macron}
\alias{GDK_dead_breve}
\alias{GDK_dead_abovedot}
\alias{GDK_dead_diaeresis}
\alias{GDK_dead_abovering}
\alias{GDK_dead_doubleacute}
\alias{GDK_dead_caron}
\alias{GDK_dead_cedilla}
\alias{GDK_dead_ogonek}
\alias{GDK_dead_iota}
\alias{GDK_dead_voiced_sound}
\alias{GDK_dead_semivoiced_sound}
\alias{GDK_dead_belowdot}
\alias{GDK_dead_hook}
\alias{GDK_dead_horn}
\alias{GDK_First_Virtual_Screen}
\alias{GDK_Prev_Virtual_Screen}
\alias{GDK_Next_Virtual_Screen}
\alias{GDK_Last_Virtual_Screen}
\alias{GDK_Terminate_Server}
\alias{GDK_AccessX_Enable}
\alias{GDK_AccessX_Feedback_Enable}
\alias{GDK_RepeatKeys_Enable}
\alias{GDK_SlowKeys_Enable}
\alias{GDK_BounceKeys_Enable}
\alias{GDK_StickyKeys_Enable}
\alias{GDK_MouseKeys_Enable}
\alias{GDK_MouseKeys_Accel_Enable}
\alias{GDK_Overlay1_Enable}
\alias{GDK_Overlay2_Enable}
\alias{GDK_AudibleBell_Enable}
\alias{GDK_Pointer_Left}
\alias{GDK_Pointer_Right}
\alias{GDK_Pointer_Up}
\alias{GDK_Pointer_Down}
\alias{GDK_Pointer_UpLeft}
\alias{GDK_Pointer_UpRight}
\alias{GDK_Pointer_DownLeft}
\alias{GDK_Pointer_DownRight}
\alias{GDK_Pointer_Button_Dflt}
\alias{GDK_Pointer_Button1}
\alias{GDK_Pointer_Button2}
\alias{GDK_Pointer_Button3}
\alias{GDK_Pointer_Button4}
\alias{GDK_Pointer_Button5}
\alias{GDK_Pointer_DblClick_Dflt}
\alias{GDK_Pointer_DblClick1}
\alias{GDK_Pointer_DblClick2}
\alias{GDK_Pointer_DblClick3}
\alias{GDK_Pointer_DblClick4}
\alias{GDK_Pointer_DblClick5}
\alias{GDK_Pointer_Drag_Dflt}
\alias{GDK_Pointer_Drag1}
\alias{GDK_Pointer_Drag2}
\alias{GDK_Pointer_Drag3}
\alias{GDK_Pointer_Drag4}
\alias{GDK_Pointer_Drag5}
\alias{GDK_Pointer_EnableKeys}
\alias{GDK_Pointer_Accelerate}
\alias{GDK_Pointer_DfltBtnNext}
\alias{GDK_Pointer_DfltBtnPrev}
\alias{GDK_3270_Duplicate}
\alias{GDK_3270_FieldMark}
\alias{GDK_3270_Right2}
\alias{GDK_3270_Left2}
\alias{GDK_3270_BackTab}
\alias{GDK_3270_EraseEOF}
\alias{GDK_3270_EraseInput}
\alias{GDK_3270_Reset}
\alias{GDK_3270_Quit}
\alias{GDK_3270_PA1}
\alias{GDK_3270_PA2}
\alias{GDK_3270_PA3}
\alias{GDK_3270_Test}
\alias{GDK_3270_Attn}
\alias{GDK_3270_CursorBlink}
\alias{GDK_3270_AltCursor}
\alias{GDK_3270_KeyClick}
\alias{GDK_3270_Jump}
\alias{GDK_3270_Ident}
\alias{GDK_3270_Rule}
\alias{GDK_3270_Copy}
\alias{GDK_3270_Play}
\alias{GDK_3270_Setup}
\alias{GDK_3270_Record}
\alias{GDK_3270_ChangeScreen}
\alias{GDK_3270_DeleteWord}
\alias{GDK_3270_ExSelect}
\alias{GDK_3270_CursorSelect}
\alias{GDK_3270_PrintScreen}
\alias{GDK_3270_Enter}
\alias{GDK_space}
\alias{GDK_exclam}
\alias{GDK_quotedbl}
\alias{GDK_numbersign}
\alias{GDK_dollar}
\alias{GDK_percent}
\alias{GDK_ampersand}
\alias{GDK_apostrophe}
\alias{GDK_quoteright}
\alias{GDK_parenleft}
\alias{GDK_parenright}
\alias{GDK_asterisk}
\alias{GDK_plus}
\alias{GDK_comma}
\alias{GDK_minus}
\alias{GDK_period}
\alias{GDK_slash}
\alias{GDK_0}
\alias{GDK_1}
\alias{GDK_2}
\alias{GDK_3}
\alias{GDK_4}
\alias{GDK_5}
\alias{GDK_6}
\alias{GDK_7}
\alias{GDK_8}
\alias{GDK_9}
\alias{GDK_colon}
\alias{GDK_semicolon}
\alias{GDK_less}
\alias{GDK_equal}
\alias{GDK_greater}
\alias{GDK_question}
\alias{GDK_at}
\alias{GDK_A}
\alias{GDK_B}
\alias{GDK_C}
\alias{GDK_D}
\alias{GDK_E}
\alias{GDK_F}
\alias{GDK_G}
\alias{GDK_H}
\alias{GDK_I}
\alias{GDK_J}
\alias{GDK_K}
\alias{GDK_L}
\alias{GDK_M}
\alias{GDK_N}
\alias{GDK_O}
\alias{GDK_P}
\alias{GDK_Q}
\alias{GDK_R}
\alias{GDK_S}
\alias{GDK_T}
\alias{GDK_U}
\alias{GDK_V}
\alias{GDK_W}
\alias{GDK_X}
\alias{GDK_Y}
\alias{GDK_Z}
\alias{GDK_bracketleft}
\alias{GDK_backslash}
\alias{GDK_bracketright}
\alias{GDK_asciicircum}
\alias{GDK_underscore}
\alias{GDK_grave}
\alias{GDK_quoteleft}
\alias{GDK_a}
\alias{GDK_b}
\alias{GDK_c}
\alias{GDK_d}
\alias{GDK_e}
\alias{GDK_f}
\alias{GDK_g}
\alias{GDK_h}
\alias{GDK_i}
\alias{GDK_j}
\alias{GDK_k}
\alias{GDK_l}
\alias{GDK_m}
\alias{GDK_n}
\alias{GDK_o}
\alias{GDK_p}
\alias{GDK_q}
\alias{GDK_r}
\alias{GDK_s}
\alias{GDK_t}
\alias{GDK_u}
\alias{GDK_v}
\alias{GDK_w}
\alias{GDK_x}
\alias{GDK_y}
\alias{GDK_z}
\alias{GDK_braceleft}
\alias{GDK_bar}
\alias{GDK_braceright}
\alias{GDK_asciitilde}
\alias{GDK_nobreakspace}
\alias{GDK_exclamdown}
\alias{GDK_cent}
\alias{GDK_sterling}
\alias{GDK_currency}
\alias{GDK_yen}
\alias{GDK_brokenbar}
\alias{GDK_section}
\alias{GDK_diaeresis}
\alias{GDK_copyright}
\alias{GDK_ordfeminine}
\alias{GDK_guillemotleft}
\alias{GDK_notsign}
\alias{GDK_hyphen}
\alias{GDK_registered}
\alias{GDK_macron}
\alias{GDK_degree}
\alias{GDK_plusminus}
\alias{GDK_twosuperior}
\alias{GDK_threesuperior}
\alias{GDK_acute}
\alias{GDK_mu}
\alias{GDK_paragraph}
\alias{GDK_periodcentered}
\alias{GDK_cedilla}
\alias{GDK_onesuperior}
\alias{GDK_masculine}
\alias{GDK_guillemotright}
\alias{GDK_onequarter}
\alias{GDK_onehalf}
\alias{GDK_threequarters}
\alias{GDK_questiondown}
\alias{GDK_Agrave}
\alias{GDK_Aacute}
\alias{GDK_Acircumflex}
\alias{GDK_Atilde}
\alias{GDK_Adiaeresis}
\alias{GDK_Aring}
\alias{GDK_AE}
\alias{GDK_Ccedilla}
\alias{GDK_Egrave}
\alias{GDK_Eacute}
\alias{GDK_Ecircumflex}
\alias{GDK_Ediaeresis}
\alias{GDK_Igrave}
\alias{GDK_Iacute}
\alias{GDK_Icircumflex}
\alias{GDK_Idiaeresis}
\alias{GDK_ETH}
\alias{GDK_Eth}
\alias{GDK_Ntilde}
\alias{GDK_Ograve}
\alias{GDK_Oacute}
\alias{GDK_Ocircumflex}
\alias{GDK_Otilde}
\alias{GDK_Odiaeresis}
\alias{GDK_multiply}
\alias{GDK_Ooblique}
\alias{GDK_Ugrave}
\alias{GDK_Uacute}
\alias{GDK_Ucircumflex}
\alias{GDK_Udiaeresis}
\alias{GDK_Yacute}
\alias{GDK_THORN}
\alias{GDK_Thorn}
\alias{GDK_ssharp}
\alias{GDK_agrave}
\alias{GDK_aacute}
\alias{GDK_acircumflex}
\alias{GDK_atilde}
\alias{GDK_adiaeresis}
\alias{GDK_aring}
\alias{GDK_ae}
\alias{GDK_ccedilla}
\alias{GDK_egrave}
\alias{GDK_eacute}
\alias{GDK_ecircumflex}
\alias{GDK_ediaeresis}
\alias{GDK_igrave}
\alias{GDK_iacute}
\alias{GDK_icircumflex}
\alias{GDK_idiaeresis}
\alias{GDK_eth}
\alias{GDK_ntilde}
\alias{GDK_ograve}
\alias{GDK_oacute}
\alias{GDK_ocircumflex}
\alias{GDK_otilde}
\alias{GDK_odiaeresis}
\alias{GDK_division}
\alias{GDK_oslash}
\alias{GDK_ugrave}
\alias{GDK_uacute}
\alias{GDK_ucircumflex}
\alias{GDK_udiaeresis}
\alias{GDK_yacute}
\alias{GDK_thorn}
\alias{GDK_ydiaeresis}
\alias{GDK_Aogonek}
\alias{GDK_breve}
\alias{GDK_Lstroke}
\alias{GDK_Lcaron}
\alias{GDK_Sacute}
\alias{GDK_Scaron}
\alias{GDK_Scedilla}
\alias{GDK_Tcaron}
\alias{GDK_Zacute}
\alias{GDK_Zcaron}
\alias{GDK_Zabovedot}
\alias{GDK_aogonek}
\alias{GDK_ogonek}
\alias{GDK_lstroke}
\alias{GDK_lcaron}
\alias{GDK_sacute}
\alias{GDK_caron}
\alias{GDK_scaron}
\alias{GDK_scedilla}
\alias{GDK_tcaron}
\alias{GDK_zacute}
\alias{GDK_doubleacute}
\alias{GDK_zcaron}
\alias{GDK_zabovedot}
\alias{GDK_Racute}
\alias{GDK_Abreve}
\alias{GDK_Lacute}
\alias{GDK_Cacute}
\alias{GDK_Ccaron}
\alias{GDK_Eogonek}
\alias{GDK_Ecaron}
\alias{GDK_Dcaron}
\alias{GDK_Dstroke}
\alias{GDK_Nacute}
\alias{GDK_Ncaron}
\alias{GDK_Odoubleacute}
\alias{GDK_Rcaron}
\alias{GDK_Uring}
\alias{GDK_Udoubleacute}
\alias{GDK_Tcedilla}
\alias{GDK_racute}
\alias{GDK_abreve}
\alias{GDK_lacute}
\alias{GDK_cacute}
\alias{GDK_ccaron}
\alias{GDK_eogonek}
\alias{GDK_ecaron}
\alias{GDK_dcaron}
\alias{GDK_dstroke}
\alias{GDK_nacute}
\alias{GDK_ncaron}
\alias{GDK_odoubleacute}
\alias{GDK_udoubleacute}
\alias{GDK_rcaron}
\alias{GDK_uring}
\alias{GDK_tcedilla}
\alias{GDK_abovedot}
\alias{GDK_Hstroke}
\alias{GDK_Hcircumflex}
\alias{GDK_Iabovedot}
\alias{GDK_Gbreve}
\alias{GDK_Jcircumflex}
\alias{GDK_hstroke}
\alias{GDK_hcircumflex}
\alias{GDK_idotless}
\alias{GDK_gbreve}
\alias{GDK_jcircumflex}
\alias{GDK_Cabovedot}
\alias{GDK_Ccircumflex}
\alias{GDK_Gabovedot}
\alias{GDK_Gcircumflex}
\alias{GDK_Ubreve}
\alias{GDK_Scircumflex}
\alias{GDK_cabovedot}
\alias{GDK_ccircumflex}
\alias{GDK_gabovedot}
\alias{GDK_gcircumflex}
\alias{GDK_ubreve}
\alias{GDK_scircumflex}
\alias{GDK_kra}
\alias{GDK_kappa}
\alias{GDK_Rcedilla}
\alias{GDK_Itilde}
\alias{GDK_Lcedilla}
\alias{GDK_Emacron}
\alias{GDK_Gcedilla}
\alias{GDK_Tslash}
\alias{GDK_rcedilla}
\alias{GDK_itilde}
\alias{GDK_lcedilla}
\alias{GDK_emacron}
\alias{GDK_gcedilla}
\alias{GDK_tslash}
\alias{GDK_ENG}
\alias{GDK_eng}
\alias{GDK_Amacron}
\alias{GDK_Iogonek}
\alias{GDK_Eabovedot}
\alias{GDK_Imacron}
\alias{GDK_Ncedilla}
\alias{GDK_Omacron}
\alias{GDK_Kcedilla}
\alias{GDK_Uogonek}
\alias{GDK_Utilde}
\alias{GDK_Umacron}
\alias{GDK_amacron}
\alias{GDK_iogonek}
\alias{GDK_eabovedot}
\alias{GDK_imacron}
\alias{GDK_ncedilla}
\alias{GDK_omacron}
\alias{GDK_kcedilla}
\alias{GDK_uogonek}
\alias{GDK_utilde}
\alias{GDK_umacron}
\alias{GDK_OE}
\alias{GDK_oe}
\alias{GDK_Ydiaeresis}
\alias{GDK_overline}
\alias{GDK_kana_fullstop}
\alias{GDK_kana_openingbracket}
\alias{GDK_kana_closingbracket}
\alias{GDK_kana_comma}
\alias{GDK_kana_conjunctive}
\alias{GDK_kana_middledot}
\alias{GDK_kana_WO}
\alias{GDK_kana_a}
\alias{GDK_kana_i}
\alias{GDK_kana_u}
\alias{GDK_kana_e}
\alias{GDK_kana_o}
\alias{GDK_kana_ya}
\alias{GDK_kana_yu}
\alias{GDK_kana_yo}
\alias{GDK_kana_tsu}
\alias{GDK_kana_tu}
\alias{GDK_prolongedsound}
\alias{GDK_kana_A}
\alias{GDK_kana_I}
\alias{GDK_kana_U}
\alias{GDK_kana_E}
\alias{GDK_kana_O}
\alias{GDK_kana_KA}
\alias{GDK_kana_KI}
\alias{GDK_kana_KU}
\alias{GDK_kana_KE}
\alias{GDK_kana_KO}
\alias{GDK_kana_SA}
\alias{GDK_kana_SHI}
\alias{GDK_kana_SU}
\alias{GDK_kana_SE}
\alias{GDK_kana_SO}
\alias{GDK_kana_TA}
\alias{GDK_kana_CHI}
\alias{GDK_kana_TI}
\alias{GDK_kana_TSU}
\alias{GDK_kana_TU}
\alias{GDK_kana_TE}
\alias{GDK_kana_TO}
\alias{GDK_kana_NA}
\alias{GDK_kana_NI}
\alias{GDK_kana_NU}
\alias{GDK_kana_NE}
\alias{GDK_kana_NO}
\alias{GDK_kana_HA}
\alias{GDK_kana_HI}
\alias{GDK_kana_FU}
\alias{GDK_kana_HU}
\alias{GDK_kana_HE}
\alias{GDK_kana_HO}
\alias{GDK_kana_MA}
\alias{GDK_kana_MI}
\alias{GDK_kana_MU}
\alias{GDK_kana_ME}
\alias{GDK_kana_MO}
\alias{GDK_kana_YA}
\alias{GDK_kana_YU}
\alias{GDK_kana_YO}
\alias{GDK_kana_RA}
\alias{GDK_kana_RI}
\alias{GDK_kana_RU}
\alias{GDK_kana_RE}
\alias{GDK_kana_RO}
\alias{GDK_kana_WA}
\alias{GDK_kana_N}
\alias{GDK_voicedsound}
\alias{GDK_semivoicedsound}
\alias{GDK_kana_switch}
\alias{GDK_Arabic_comma}
\alias{GDK_Arabic_semicolon}
\alias{GDK_Arabic_question_mark}
\alias{GDK_Arabic_hamza}
\alias{GDK_Arabic_maddaonalef}
\alias{GDK_Arabic_hamzaonalef}
\alias{GDK_Arabic_hamzaonwaw}
\alias{GDK_Arabic_hamzaunderalef}
\alias{GDK_Arabic_hamzaonyeh}
\alias{GDK_Arabic_alef}
\alias{GDK_Arabic_beh}
\alias{GDK_Arabic_tehmarbuta}
\alias{GDK_Arabic_teh}
\alias{GDK_Arabic_theh}
\alias{GDK_Arabic_jeem}
\alias{GDK_Arabic_hah}
\alias{GDK_Arabic_khah}
\alias{GDK_Arabic_dal}
\alias{GDK_Arabic_thal}
\alias{GDK_Arabic_ra}
\alias{GDK_Arabic_zain}
\alias{GDK_Arabic_seen}
\alias{GDK_Arabic_sheen}
\alias{GDK_Arabic_sad}
\alias{GDK_Arabic_dad}
\alias{GDK_Arabic_tah}
\alias{GDK_Arabic_zah}
\alias{GDK_Arabic_ain}
\alias{GDK_Arabic_ghain}
\alias{GDK_Arabic_tatweel}
\alias{GDK_Arabic_feh}
\alias{GDK_Arabic_qaf}
\alias{GDK_Arabic_kaf}
\alias{GDK_Arabic_lam}
\alias{GDK_Arabic_meem}
\alias{GDK_Arabic_noon}
\alias{GDK_Arabic_ha}
\alias{GDK_Arabic_heh}
\alias{GDK_Arabic_waw}
\alias{GDK_Arabic_alefmaksura}
\alias{GDK_Arabic_yeh}
\alias{GDK_Arabic_fathatan}
\alias{GDK_Arabic_dammatan}
\alias{GDK_Arabic_kasratan}
\alias{GDK_Arabic_fatha}
\alias{GDK_Arabic_damma}
\alias{GDK_Arabic_kasra}
\alias{GDK_Arabic_shadda}
\alias{GDK_Arabic_sukun}
\alias{GDK_Arabic_switch}
\alias{GDK_Serbian_dje}
\alias{GDK_Macedonia_gje}
\alias{GDK_Cyrillic_io}
\alias{GDK_Ukrainian_ie}
\alias{GDK_Ukranian_je}
\alias{GDK_Macedonia_dse}
\alias{GDK_Ukrainian_i}
\alias{GDK_Ukranian_i}
\alias{GDK_Ukrainian_yi}
\alias{GDK_Ukranian_yi}
\alias{GDK_Cyrillic_je}
\alias{GDK_Serbian_je}
\alias{GDK_Cyrillic_lje}
\alias{GDK_Serbian_lje}
\alias{GDK_Cyrillic_nje}
\alias{GDK_Serbian_nje}
\alias{GDK_Serbian_tshe}
\alias{GDK_Macedonia_kje}
\alias{GDK_Ukrainian_ghe_with_upturn}
\alias{GDK_Byelorussian_shortu}
\alias{GDK_Cyrillic_dzhe}
\alias{GDK_Serbian_dze}
\alias{GDK_numerosign}
\alias{GDK_Serbian_DJE}
\alias{GDK_Macedonia_GJE}
\alias{GDK_Cyrillic_IO}
\alias{GDK_Ukrainian_IE}
\alias{GDK_Ukranian_JE}
\alias{GDK_Macedonia_DSE}
\alias{GDK_Ukrainian_I}
\alias{GDK_Ukranian_I}
\alias{GDK_Ukrainian_YI}
\alias{GDK_Ukranian_YI}
\alias{GDK_Cyrillic_JE}
\alias{GDK_Serbian_JE}
\alias{GDK_Cyrillic_LJE}
\alias{GDK_Serbian_LJE}
\alias{GDK_Cyrillic_NJE}
\alias{GDK_Serbian_NJE}
\alias{GDK_Serbian_TSHE}
\alias{GDK_Macedonia_KJE}
\alias{GDK_Ukrainian_GHE_WITH_UPTURN}
\alias{GDK_Byelorussian_SHORTU}
\alias{GDK_Cyrillic_DZHE}
\alias{GDK_Serbian_DZE}
\alias{GDK_Cyrillic_yu}
\alias{GDK_Cyrillic_a}
\alias{GDK_Cyrillic_be}
\alias{GDK_Cyrillic_tse}
\alias{GDK_Cyrillic_de}
\alias{GDK_Cyrillic_ie}
\alias{GDK_Cyrillic_ef}
\alias{GDK_Cyrillic_ghe}
\alias{GDK_Cyrillic_ha}
\alias{GDK_Cyrillic_i}
\alias{GDK_Cyrillic_shorti}
\alias{GDK_Cyrillic_ka}
\alias{GDK_Cyrillic_el}
\alias{GDK_Cyrillic_em}
\alias{GDK_Cyrillic_en}
\alias{GDK_Cyrillic_o}
\alias{GDK_Cyrillic_pe}
\alias{GDK_Cyrillic_ya}
\alias{GDK_Cyrillic_er}
\alias{GDK_Cyrillic_es}
\alias{GDK_Cyrillic_te}
\alias{GDK_Cyrillic_u}
\alias{GDK_Cyrillic_zhe}
\alias{GDK_Cyrillic_ve}
\alias{GDK_Cyrillic_softsign}
\alias{GDK_Cyrillic_yeru}
\alias{GDK_Cyrillic_ze}
\alias{GDK_Cyrillic_sha}
\alias{GDK_Cyrillic_e}
\alias{GDK_Cyrillic_shcha}
\alias{GDK_Cyrillic_che}
\alias{GDK_Cyrillic_hardsign}
\alias{GDK_Cyrillic_YU}
\alias{GDK_Cyrillic_A}
\alias{GDK_Cyrillic_BE}
\alias{GDK_Cyrillic_TSE}
\alias{GDK_Cyrillic_DE}
\alias{GDK_Cyrillic_IE}
\alias{GDK_Cyrillic_EF}
\alias{GDK_Cyrillic_GHE}
\alias{GDK_Cyrillic_HA}
\alias{GDK_Cyrillic_I}
\alias{GDK_Cyrillic_SHORTI}
\alias{GDK_Cyrillic_KA}
\alias{GDK_Cyrillic_EL}
\alias{GDK_Cyrillic_EM}
\alias{GDK_Cyrillic_EN}
\alias{GDK_Cyrillic_O}
\alias{GDK_Cyrillic_PE}
\alias{GDK_Cyrillic_YA}
\alias{GDK_Cyrillic_ER}
\alias{GDK_Cyrillic_ES}
\alias{GDK_Cyrillic_TE}
\alias{GDK_Cyrillic_U}
\alias{GDK_Cyrillic_ZHE}
\alias{GDK_Cyrillic_VE}
\alias{GDK_Cyrillic_SOFTSIGN}
\alias{GDK_Cyrillic_YERU}
\alias{GDK_Cyrillic_ZE}
\alias{GDK_Cyrillic_SHA}
\alias{GDK_Cyrillic_E}
\alias{GDK_Cyrillic_SHCHA}
\alias{GDK_Cyrillic_CHE}
\alias{GDK_Cyrillic_HARDSIGN}
\alias{GDK_Greek_ALPHAaccent}
\alias{GDK_Greek_EPSILONaccent}
\alias{GDK_Greek_ETAaccent}
\alias{GDK_Greek_IOTAaccent}
\alias{GDK_Greek_IOTAdieresis}
\alias{GDK_Greek_IOTAdiaeresis}
\alias{GDK_Greek_OMICRONaccent}
\alias{GDK_Greek_UPSILONaccent}
\alias{GDK_Greek_UPSILONdieresis}
\alias{GDK_Greek_OMEGAaccent}
\alias{GDK_Greek_accentdieresis}
\alias{GDK_Greek_horizbar}
\alias{GDK_Greek_alphaaccent}
\alias{GDK_Greek_epsilonaccent}
\alias{GDK_Greek_etaaccent}
\alias{GDK_Greek_iotaaccent}
\alias{GDK_Greek_iotadieresis}
\alias{GDK_Greek_iotaaccentdieresis}
\alias{GDK_Greek_omicronaccent}
\alias{GDK_Greek_upsilonaccent}
\alias{GDK_Greek_upsilondieresis}
\alias{GDK_Greek_upsilonaccentdieresis}
\alias{GDK_Greek_omegaaccent}
\alias{GDK_Greek_ALPHA}
\alias{GDK_Greek_BETA}
\alias{GDK_Greek_GAMMA}
\alias{GDK_Greek_DELTA}
\alias{GDK_Greek_EPSILON}
\alias{GDK_Greek_ZETA}
\alias{GDK_Greek_ETA}
\alias{GDK_Greek_THETA}
\alias{GDK_Greek_IOTA}
\alias{GDK_Greek_KAPPA}
\alias{GDK_Greek_LAMDA}
\alias{GDK_Greek_LAMBDA}
\alias{GDK_Greek_MU}
\alias{GDK_Greek_NU}
\alias{GDK_Greek_XI}
\alias{GDK_Greek_OMICRON}
\alias{GDK_Greek_PI}
\alias{GDK_Greek_RHO}
\alias{GDK_Greek_SIGMA}
\alias{GDK_Greek_TAU}
\alias{GDK_Greek_UPSILON}
\alias{GDK_Greek_PHI}
\alias{GDK_Greek_CHI}
\alias{GDK_Greek_PSI}
\alias{GDK_Greek_OMEGA}
\alias{GDK_Greek_alpha}
\alias{GDK_Greek_beta}
\alias{GDK_Greek_gamma}
\alias{GDK_Greek_delta}
\alias{GDK_Greek_epsilon}
\alias{GDK_Greek_zeta}
\alias{GDK_Greek_eta}
\alias{GDK_Greek_theta}
\alias{GDK_Greek_iota}
\alias{GDK_Greek_kappa}
\alias{GDK_Greek_lamda}
\alias{GDK_Greek_lambda}
\alias{GDK_Greek_mu}
\alias{GDK_Greek_nu}
\alias{GDK_Greek_xi}
\alias{GDK_Greek_omicron}
\alias{GDK_Greek_pi}
\alias{GDK_Greek_rho}
\alias{GDK_Greek_sigma}
\alias{GDK_Greek_finalsmallsigma}
\alias{GDK_Greek_tau}
\alias{GDK_Greek_upsilon}
\alias{GDK_Greek_phi}
\alias{GDK_Greek_chi}
\alias{GDK_Greek_psi}
\alias{GDK_Greek_omega}
\alias{GDK_Greek_switch}
\alias{GDK_leftradical}
\alias{GDK_topleftradical}
\alias{GDK_horizconnector}
\alias{GDK_topintegral}
\alias{GDK_botintegral}
\alias{GDK_vertconnector}
\alias{GDK_topleftsqbracket}
\alias{GDK_botleftsqbracket}
\alias{GDK_toprightsqbracket}
\alias{GDK_botrightsqbracket}
\alias{GDK_topleftparens}
\alias{GDK_botleftparens}
\alias{GDK_toprightparens}
\alias{GDK_botrightparens}
\alias{GDK_leftmiddlecurlybrace}
\alias{GDK_rightmiddlecurlybrace}
\alias{GDK_topleftsummation}
\alias{GDK_botleftsummation}
\alias{GDK_topvertsummationconnector}
\alias{GDK_botvertsummationconnector}
\alias{GDK_toprightsummation}
\alias{GDK_botrightsummation}
\alias{GDK_rightmiddlesummation}
\alias{GDK_lessthanequal}
\alias{GDK_notequal}
\alias{GDK_greaterthanequal}
\alias{GDK_integral}
\alias{GDK_therefore}
\alias{GDK_variation}
\alias{GDK_infinity}
\alias{GDK_nabla}
\alias{GDK_approximate}
\alias{GDK_similarequal}
\alias{GDK_ifonlyif}
\alias{GDK_implies}
\alias{GDK_identical}
\alias{GDK_radical}
\alias{GDK_includedin}
\alias{GDK_includes}
\alias{GDK_intersection}
\alias{GDK_union}
\alias{GDK_logicaland}
\alias{GDK_logicalor}
\alias{GDK_partialderivative}
\alias{GDK_function}
\alias{GDK_leftarrow}
\alias{GDK_uparrow}
\alias{GDK_rightarrow}
\alias{GDK_downarrow}
\alias{GDK_blank}
\alias{GDK_soliddiamond}
\alias{GDK_checkerboard}
\alias{GDK_ht}
\alias{GDK_ff}
\alias{GDK_cr}
\alias{GDK_lf}
\alias{GDK_nl}
\alias{GDK_vt}
\alias{GDK_lowrightcorner}
\alias{GDK_uprightcorner}
\alias{GDK_upleftcorner}
\alias{GDK_lowleftcorner}
\alias{GDK_crossinglines}
\alias{GDK_horizlinescan1}
\alias{GDK_horizlinescan3}
\alias{GDK_horizlinescan5}
\alias{GDK_horizlinescan7}
\alias{GDK_horizlinescan9}
\alias{GDK_leftt}
\alias{GDK_rightt}
\alias{GDK_bott}
\alias{GDK_topt}
\alias{GDK_vertbar}
\alias{GDK_emspace}
\alias{GDK_enspace}
\alias{GDK_em3space}
\alias{GDK_em4space}
\alias{GDK_digitspace}
\alias{GDK_punctspace}
\alias{GDK_thinspace}
\alias{GDK_hairspace}
\alias{GDK_emdash}
\alias{GDK_endash}
\alias{GDK_signifblank}
\alias{GDK_ellipsis}
\alias{GDK_doubbaselinedot}
\alias{GDK_onethird}
\alias{GDK_twothirds}
\alias{GDK_onefifth}
\alias{GDK_twofifths}
\alias{GDK_threefifths}
\alias{GDK_fourfifths}
\alias{GDK_onesixth}
\alias{GDK_fivesixths}
\alias{GDK_careof}
\alias{GDK_figdash}
\alias{GDK_leftanglebracket}
\alias{GDK_decimalpoint}
\alias{GDK_rightanglebracket}
\alias{GDK_marker}
\alias{GDK_oneeighth}
\alias{GDK_threeeighths}
\alias{GDK_fiveeighths}
\alias{GDK_seveneighths}
\alias{GDK_trademark}
\alias{GDK_signaturemark}
\alias{GDK_trademarkincircle}
\alias{GDK_leftopentriangle}
\alias{GDK_rightopentriangle}
\alias{GDK_emopencircle}
\alias{GDK_emopenrectangle}
\alias{GDK_leftsinglequotemark}
\alias{GDK_rightsinglequotemark}
\alias{GDK_leftdoublequotemark}
\alias{GDK_rightdoublequotemark}
\alias{GDK_prescription}
\alias{GDK_minutes}
\alias{GDK_seconds}
\alias{GDK_latincross}
\alias{GDK_hexagram}
\alias{GDK_filledrectbullet}
\alias{GDK_filledlefttribullet}
\alias{GDK_filledrighttribullet}
\alias{GDK_emfilledcircle}
\alias{GDK_emfilledrect}
\alias{GDK_enopencircbullet}
\alias{GDK_enopensquarebullet}
\alias{GDK_openrectbullet}
\alias{GDK_opentribulletup}
\alias{GDK_opentribulletdown}
\alias{GDK_openstar}
\alias{GDK_enfilledcircbullet}
\alias{GDK_enfilledsqbullet}
\alias{GDK_filledtribulletup}
\alias{GDK_filledtribulletdown}
\alias{GDK_leftpointer}
\alias{GDK_rightpointer}
\alias{GDK_club}
\alias{GDK_diamond}
\alias{GDK_heart}
\alias{GDK_maltesecross}
\alias{GDK_dagger}
\alias{GDK_doubledagger}
\alias{GDK_checkmark}
\alias{GDK_ballotcross}
\alias{GDK_musicalsharp}
\alias{GDK_musicalflat}
\alias{GDK_malesymbol}
\alias{GDK_femalesymbol}
\alias{GDK_telephone}
\alias{GDK_telephonerecorder}
\alias{GDK_phonographcopyright}
\alias{GDK_caret}
\alias{GDK_singlelowquotemark}
\alias{GDK_doublelowquotemark}
\alias{GDK_cursor}
\alias{GDK_leftcaret}
\alias{GDK_rightcaret}
\alias{GDK_downcaret}
\alias{GDK_upcaret}
\alias{GDK_overbar}
\alias{GDK_downtack}
\alias{GDK_upshoe}
\alias{GDK_downstile}
\alias{GDK_underbar}
\alias{GDK_jot}
\alias{GDK_quad}
\alias{GDK_uptack}
\alias{GDK_circle}
\alias{GDK_upstile}
\alias{GDK_downshoe}
\alias{GDK_rightshoe}
\alias{GDK_leftshoe}
\alias{GDK_lefttack}
\alias{GDK_righttack}
\alias{GDK_hebrew_doublelowline}
\alias{GDK_hebrew_aleph}
\alias{GDK_hebrew_bet}
\alias{GDK_hebrew_beth}
\alias{GDK_hebrew_gimel}
\alias{GDK_hebrew_gimmel}
\alias{GDK_hebrew_dalet}
\alias{GDK_hebrew_daleth}
\alias{GDK_hebrew_he}
\alias{GDK_hebrew_waw}
\alias{GDK_hebrew_zain}
\alias{GDK_hebrew_zayin}
\alias{GDK_hebrew_chet}
\alias{GDK_hebrew_het}
\alias{GDK_hebrew_tet}
\alias{GDK_hebrew_teth}
\alias{GDK_hebrew_yod}
\alias{GDK_hebrew_finalkaph}
\alias{GDK_hebrew_kaph}
\alias{GDK_hebrew_lamed}
\alias{GDK_hebrew_finalmem}
\alias{GDK_hebrew_mem}
\alias{GDK_hebrew_finalnun}
\alias{GDK_hebrew_nun}
\alias{GDK_hebrew_samech}
\alias{GDK_hebrew_samekh}
\alias{GDK_hebrew_ayin}
\alias{GDK_hebrew_finalpe}
\alias{GDK_hebrew_pe}
\alias{GDK_hebrew_finalzade}
\alias{GDK_hebrew_finalzadi}
\alias{GDK_hebrew_zade}
\alias{GDK_hebrew_zadi}
\alias{GDK_hebrew_qoph}
\alias{GDK_hebrew_kuf}
\alias{GDK_hebrew_resh}
\alias{GDK_hebrew_shin}
\alias{GDK_hebrew_taw}
\alias{GDK_hebrew_taf}
\alias{GDK_Hebrew_switch}
\alias{GDK_Thai_kokai}
\alias{GDK_Thai_khokhai}
\alias{GDK_Thai_khokhuat}
\alias{GDK_Thai_khokhwai}
\alias{GDK_Thai_khokhon}
\alias{GDK_Thai_khorakhang}
\alias{GDK_Thai_ngongu}
\alias{GDK_Thai_chochan}
\alias{GDK_Thai_choching}
\alias{GDK_Thai_chochang}
\alias{GDK_Thai_soso}
\alias{GDK_Thai_chochoe}
\alias{GDK_Thai_yoying}
\alias{GDK_Thai_dochada}
\alias{GDK_Thai_topatak}
\alias{GDK_Thai_thothan}
\alias{GDK_Thai_thonangmontho}
\alias{GDK_Thai_thophuthao}
\alias{GDK_Thai_nonen}
\alias{GDK_Thai_dodek}
\alias{GDK_Thai_totao}
\alias{GDK_Thai_thothung}
\alias{GDK_Thai_thothahan}
\alias{GDK_Thai_thothong}
\alias{GDK_Thai_nonu}
\alias{GDK_Thai_bobaimai}
\alias{GDK_Thai_popla}
\alias{GDK_Thai_phophung}
\alias{GDK_Thai_fofa}
\alias{GDK_Thai_phophan}
\alias{GDK_Thai_fofan}
\alias{GDK_Thai_phosamphao}
\alias{GDK_Thai_moma}
\alias{GDK_Thai_yoyak}
\alias{GDK_Thai_rorua}
\alias{GDK_Thai_ru}
\alias{GDK_Thai_loling}
\alias{GDK_Thai_lu}
\alias{GDK_Thai_wowaen}
\alias{GDK_Thai_sosala}
\alias{GDK_Thai_sorusi}
\alias{GDK_Thai_sosua}
\alias{GDK_Thai_hohip}
\alias{GDK_Thai_lochula}
\alias{GDK_Thai_oang}
\alias{GDK_Thai_honokhuk}
\alias{GDK_Thai_paiyannoi}
\alias{GDK_Thai_saraa}
\alias{GDK_Thai_maihanakat}
\alias{GDK_Thai_saraaa}
\alias{GDK_Thai_saraam}
\alias{GDK_Thai_sarai}
\alias{GDK_Thai_saraii}
\alias{GDK_Thai_saraue}
\alias{GDK_Thai_sarauee}
\alias{GDK_Thai_sarau}
\alias{GDK_Thai_sarauu}
\alias{GDK_Thai_phinthu}
\alias{GDK_Thai_maihanakat_maitho}
\alias{GDK_Thai_baht}
\alias{GDK_Thai_sarae}
\alias{GDK_Thai_saraae}
\alias{GDK_Thai_sarao}
\alias{GDK_Thai_saraaimaimuan}
\alias{GDK_Thai_saraaimaimalai}
\alias{GDK_Thai_lakkhangyao}
\alias{GDK_Thai_maiyamok}
\alias{GDK_Thai_maitaikhu}
\alias{GDK_Thai_maiek}
\alias{GDK_Thai_maitho}
\alias{GDK_Thai_maitri}
\alias{GDK_Thai_maichattawa}
\alias{GDK_Thai_thanthakhat}
\alias{GDK_Thai_nikhahit}
\alias{GDK_Thai_leksun}
\alias{GDK_Thai_leknung}
\alias{GDK_Thai_leksong}
\alias{GDK_Thai_leksam}
\alias{GDK_Thai_leksi}
\alias{GDK_Thai_lekha}
\alias{GDK_Thai_lekhok}
\alias{GDK_Thai_lekchet}
\alias{GDK_Thai_lekpaet}
\alias{GDK_Thai_lekkao}
\alias{GDK_Hangul}
\alias{GDK_Hangul_Start}
\alias{GDK_Hangul_End}
\alias{GDK_Hangul_Hanja}
\alias{GDK_Hangul_Jamo}
\alias{GDK_Hangul_Romaja}
\alias{GDK_Hangul_Codeinput}
\alias{GDK_Hangul_Jeonja}
\alias{GDK_Hangul_Banja}
\alias{GDK_Hangul_PreHanja}
\alias{GDK_Hangul_PostHanja}
\alias{GDK_Hangul_SingleCandidate}
\alias{GDK_Hangul_MultipleCandidate}
\alias{GDK_Hangul_PreviousCandidate}
\alias{GDK_Hangul_Special}
\alias{GDK_Hangul_switch}
\alias{GDK_Hangul_Kiyeog}
\alias{GDK_Hangul_SsangKiyeog}
\alias{GDK_Hangul_KiyeogSios}
\alias{GDK_Hangul_Nieun}
\alias{GDK_Hangul_NieunJieuj}
\alias{GDK_Hangul_NieunHieuh}
\alias{GDK_Hangul_Dikeud}
\alias{GDK_Hangul_SsangDikeud}
\alias{GDK_Hangul_Rieul}
\alias{GDK_Hangul_RieulKiyeog}
\alias{GDK_Hangul_RieulMieum}
\alias{GDK_Hangul_RieulPieub}
\alias{GDK_Hangul_RieulSios}
\alias{GDK_Hangul_RieulTieut}
\alias{GDK_Hangul_RieulPhieuf}
\alias{GDK_Hangul_RieulHieuh}
\alias{GDK_Hangul_Mieum}
\alias{GDK_Hangul_Pieub}
\alias{GDK_Hangul_SsangPieub}
\alias{GDK_Hangul_PieubSios}
\alias{GDK_Hangul_Sios}
\alias{GDK_Hangul_SsangSios}
\alias{GDK_Hangul_Ieung}
\alias{GDK_Hangul_Jieuj}
\alias{GDK_Hangul_SsangJieuj}
\alias{GDK_Hangul_Cieuc}
\alias{GDK_Hangul_Khieuq}
\alias{GDK_Hangul_Tieut}
\alias{GDK_Hangul_Phieuf}
\alias{GDK_Hangul_Hieuh}
\alias{GDK_Hangul_A}
\alias{GDK_Hangul_AE}
\alias{GDK_Hangul_YA}
\alias{GDK_Hangul_YAE}
\alias{GDK_Hangul_EO}
\alias{GDK_Hangul_E}
\alias{GDK_Hangul_YEO}
\alias{GDK_Hangul_YE}
\alias{GDK_Hangul_O}
\alias{GDK_Hangul_WA}
\alias{GDK_Hangul_WAE}
\alias{GDK_Hangul_OE}
\alias{GDK_Hangul_YO}
\alias{GDK_Hangul_U}
\alias{GDK_Hangul_WEO}
\alias{GDK_Hangul_WE}
\alias{GDK_Hangul_WI}
\alias{GDK_Hangul_YU}
\alias{GDK_Hangul_EU}
\alias{GDK_Hangul_YI}
\alias{GDK_Hangul_I}
\alias{GDK_Hangul_J_Kiyeog}
\alias{GDK_Hangul_J_SsangKiyeog}
\alias{GDK_Hangul_J_KiyeogSios}
\alias{GDK_Hangul_J_Nieun}
\alias{GDK_Hangul_J_NieunJieuj}
\alias{GDK_Hangul_J_NieunHieuh}
\alias{GDK_Hangul_J_Dikeud}
\alias{GDK_Hangul_J_Rieul}
\alias{GDK_Hangul_J_RieulKiyeog}
\alias{GDK_Hangul_J_RieulMieum}
\alias{GDK_Hangul_J_RieulPieub}
\alias{GDK_Hangul_J_RieulSios}
\alias{GDK_Hangul_J_RieulTieut}
\alias{GDK_Hangul_J_RieulPhieuf}
\alias{GDK_Hangul_J_RieulHieuh}
\alias{GDK_Hangul_J_Mieum}
\alias{GDK_Hangul_J_Pieub}
\alias{GDK_Hangul_J_PieubSios}
\alias{GDK_Hangul_J_Sios}
\alias{GDK_Hangul_J_SsangSios}
\alias{GDK_Hangul_J_Ieung}
\alias{GDK_Hangul_J_Jieuj}
\alias{GDK_Hangul_J_Cieuc}
\alias{GDK_Hangul_J_Khieuq}
\alias{GDK_Hangul_J_Tieut}
\alias{GDK_Hangul_J_Phieuf}
\alias{GDK_Hangul_J_Hieuh}
\alias{GDK_Hangul_RieulYeorinHieuh}
\alias{GDK_Hangul_SunkyeongeumMieum}
\alias{GDK_Hangul_SunkyeongeumPieub}
\alias{GDK_Hangul_PanSios}
\alias{GDK_Hangul_KkogjiDalrinIeung}
\alias{GDK_Hangul_SunkyeongeumPhieuf}
\alias{GDK_Hangul_YeorinHieuh}
\alias{GDK_Hangul_AraeA}
\alias{GDK_Hangul_AraeAE}
\alias{GDK_Hangul_J_PanSios}
\alias{GDK_Hangul_J_KkogjiDalrinIeung}
\alias{GDK_Hangul_J_YeorinHieuh}
\alias{GDK_Korean_Won}
\alias{GDK_EcuSign}
\alias{GDK_ColonSign}
\alias{GDK_CruzeiroSign}
\alias{GDK_FFrancSign}
\alias{GDK_LiraSign}
\alias{GDK_MillSign}
\alias{GDK_NairaSign}
\alias{GDK_PesetaSign}
\alias{GDK_RupeeSign}
\alias{GDK_WonSign}
\alias{GDK_NewSheqelSign}
\alias{GDK_DongSign}
\alias{GDK_EuroSign}
\name{gdkKeySyms}
\title{GDK Key Constants}
\description{This is a GDK key constant. Use it for identifying the key pressed in a \code{\link{GdkEventKey}}.}
\author{Michael Lawrence}
\keyword{internal}
\keyword{interface}
