\alias{gdkKeyvalIsLower}
\name{gdkKeyvalIsLower}
\title{gdkKeyvalIsLower}
\description{Returns \code{TRUE} if the given key value is in lower case.}
\usage{gdkKeyvalIsLower(keyval)}
\arguments{\item{\code{keyval}}{[numeric] a key value.}}
\value{[logical] \code{TRUE} if \code{keyval} is in lower case, or if \code{keyval} is not subject to
case conversion.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
