\alias{gdkKeyvalIsUpper}
\name{gdkKeyvalIsUpper}
\title{gdkKeyvalIsUpper}
\description{Returns \code{TRUE} if the given key value is in upper case.}
\usage{gdkKeyvalIsUpper(keyval)}
\arguments{\item{\verb{keyval}}{a key value.}}
\value{[logical] \code{TRUE} if \code{keyval} is in upper case, or if \code{keyval} is not subject to
case conversion.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
