\alias{gdkKeyvalToLower}
\name{gdkKeyvalToLower}
\title{gdkKeyvalToLower}
\description{Converts a key value to lower case, if applicable.}
\usage{gdkKeyvalToLower(keyval)}
\arguments{\item{\verb{keyval}}{a key value.}}
\value{[numeric] the lower case form of \code{keyval}, or \code{keyval} itself if it is already
in lower case or it is not subject to case conversion.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
