\alias{gdkKeyvalToUnicode}
\name{gdkKeyvalToUnicode}
\title{gdkKeyvalToUnicode}
\description{Convert from a GDK key symbol to the corresponding ISO10646 (Unicode)
character.}
\usage{gdkKeyvalToUnicode(keyval)}
\arguments{\item{\verb{keyval}}{a GDK key symbol}}
\value{[numeric] the corresponding unicode character, or 0 if there
is no corresponding character.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
