\alias{gdkKeyvalToUpper}
\name{gdkKeyvalToUpper}
\title{gdkKeyvalToUpper}
\description{Converts a key value to upper case, if applicable.}
\usage{gdkKeyvalToUpper(keyval)}
\arguments{\item{\verb{keyval}}{a key value.}}
\value{[numeric] the upper case form of \code{keyval}, or \code{keyval} itself if it is already
in upper case or it is not subject to case conversion.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
