\alias{gdkListVisuals}
\name{gdkListVisuals}
\title{gdkListVisuals}
\description{Lists the available visuals for the default screen.
(See \code{\link{gdkScreenListVisuals}})
A visual describes a hardware image data format.
For example, a visual might support 24-bit color, or 8-bit color,
and might expect pixels to be in a certain format.}
\usage{gdkListVisuals()}
\details{Call \code{g\List\Free()} on the return value when you're finished with it.  }
\value{[list] .}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
