\alias{gdkPangoAttrEmbossedNew}
\name{gdkPangoAttrEmbossedNew}
\title{gdkPangoAttrEmbossedNew}
\description{Creates a new attribute flagging a region as embossed or not.}
\usage{gdkPangoAttrEmbossedNew(embossed)}
\arguments{\item{\code{embossed}}{[logical]  if the region should be embossed}}
\value{[\code{\link{PangoAttribute}}]  new \code{\link{PangoAttribute}}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
