\alias{gdkPangoRendererNew}
\name{gdkPangoRendererNew}
\title{gdkPangoRendererNew}
\description{Creates a new \verb{\link{PangoRenderer}} for \code{screen}. Normally you can use the
results of \code{\link{gdkPangoRendererGetDefault}} rather than creating a new
renderer.}
\usage{gdkPangoRendererNew(screen)}
\arguments{\item{\verb{screen}}{[\verb{\link{GdkScreen}}]  a \verb{\link{GdkScreen}}}}
\details{  Since  2.6}
\value{[\verb{\link{PangoRenderer}}]  a newly created \verb{\link{PangoRenderer}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
