\alias{gdkPixbufAnimationGetWidth}
\name{gdkPixbufAnimationGetWidth}
\title{gdkPixbufAnimationGetWidth}
\description{Queries the width of the bounding box of a pixbuf animation.}
\usage{gdkPixbufAnimationGetWidth(object)}
\arguments{\item{\verb{object}}{An animation.}}
\value{[integer] Width of the bounding box of the animation.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
