\alias{gdkPixbufCopy}
\name{gdkPixbufCopy}
\title{gdkPixbufCopy}
\description{Creates a new \verb{\link{GdkPixbuf}} with a copy of the information in the specified
\code{pixbuf}.}
\usage{gdkPixbufCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkPixbuf}}]  A pixbuf.}}
\value{[\verb{\link{GdkPixbuf}}]  A newly-created pixbuf with a reference count of 1, or \code{NULL} if
not enough memory could be allocated.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
