\alias{gdkPixbufFlip}
\name{gdkPixbufFlip}
\title{gdkPixbufFlip}
\description{Flips a pixbuf horizontally or vertically and returns the
result in a new pixbuf.}
\usage{gdkPixbufFlip(object, horizontal)}
\arguments{
\item{\code{object}}{[\code{\link{GdkPixbuf}}]  a \code{\link{GdkPixbuf}}}
\item{\code{horizontal}}{[logical]  \code{TRUE} to flip horizontally, \code{FALSE} to flip vertically}
}
\details{  Since  2.6}
\value{[\code{\link{GdkPixbuf}}]  a new pixbuf.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
