\alias{gdkPixbufFormatGetDescription}
\name{gdkPixbufFormatGetDescription}
\title{gdkPixbufFormatGetDescription}
\description{Returns a description of the format.}
\usage{gdkPixbufFormatGetDescription(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkPixbufFormat}}}}
\details{Since 2.2}
\value{[character] a description of the format.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
