\alias{gdkPixbufFormatGetMimeTypes}
\name{gdkPixbufFormatGetMimeTypes}
\title{gdkPixbufFormatGetMimeTypes}
\description{Returns the mime types supported by the format.}
\usage{gdkPixbufFormatGetMimeTypes(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkPixbufFormat}}]  a \code{\link{GdkPixbufFormat}}}}
\details{  Since  2.2}
\value{[character] .}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
