\alias{gdkPixbufFormatGetName}
\name{gdkPixbufFormatGetName}
\title{gdkPixbufFormatGetName}
\description{Returns the name of the format.}
\usage{gdkPixbufFormatGetName(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkPixbufFormat}}]  a \verb{\link{GdkPixbufFormat}}}}
\details{  Since  2.2}
\value{[character]  the name of the format. }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
