\alias{gdkPixbufFormatIsDisabled}
\name{gdkPixbufFormatIsDisabled}
\title{gdkPixbufFormatIsDisabled}
\description{Returns whether this image format is disabled. See
\code{\link{gdkPixbufFormatSetDisabled}}.}
\usage{gdkPixbufFormatIsDisabled(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkPixbufFormat}}}}
\details{Since 2.6}
\value{[logical] whether this image format is disabled.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
