\alias{gdkPixbufFormatIsWritable}
\name{gdkPixbufFormatIsWritable}
\title{gdkPixbufFormatIsWritable}
\description{Returns whether pixbufs can be saved in the given format.}
\usage{gdkPixbufFormatIsWritable(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkPixbufFormat}}]  a \code{\link{GdkPixbufFormat}}}}
\details{  Since  2.2}
\value{[logical]  whether pixbufs can be saved in the given format.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
