\alias{gdkPixbufGetBitsPerSample}
\name{gdkPixbufGetBitsPerSample}
\title{gdkPixbufGetBitsPerSample}
\description{Queries the number of bits per color sample in a pixbuf.}
\usage{gdkPixbufGetBitsPerSample(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkPixbuf}}]  A pixbuf.}}
\value{[integer]  Number of bits per color sample.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
