\alias{gdkPixbufGetColorspace}
\name{gdkPixbufGetColorspace}
\title{gdkPixbufGetColorspace}
\description{Queries the color space of a pixbuf.}
\usage{gdkPixbufGetColorspace(object)}
\arguments{\item{\verb{object}}{A pixbuf.}}
\value{[\code{\link{GdkColorspace}}] Color space.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
