\alias{gdkPixbufGetHeight}
\name{gdkPixbufGetHeight}
\title{gdkPixbufGetHeight}
\description{Queries the height of a pixbuf.}
\usage{gdkPixbufGetHeight(object)}
\arguments{\item{\verb{object}}{A pixbuf.}}
\value{[integer] Height in pixels.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
