\alias{gdkPixbufGetNChannels}
\name{gdkPixbufGetNChannels}
\title{gdkPixbufGetNChannels}
\description{Queries the number of channels of a pixbuf.}
\usage{gdkPixbufGetNChannels(object)}
\arguments{\item{\verb{object}}{A pixbuf.}}
\value{[integer] Number of channels.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
