\alias{gdkPixbufGetOption}
\name{gdkPixbufGetOption}
\title{gdkPixbufGetOption}
\description{Looks up \code{key} in the list of options that may have been attached to the
\code{pixbuf} when it was loaded.}
\usage{gdkPixbufGetOption(object, key)}
\arguments{
\item{\code{object}}{[\code{\link{GdkPixbuf}}]  a \code{\link{GdkPixbuf}}}
\item{\code{key}}{[character]  a string.}
}
\value{[character]  the value associated with \code{key}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
