\alias{gdkPixbufGetPixels}
\name{gdkPixbufGetPixels}
\title{gdkPixbufGetPixels}
\description{Queries a pointer to the pixel data of a pixbuf.}
\usage{gdkPixbufGetPixels(object)}
\arguments{\item{\verb{object}}{A pixbuf.}}
\value{[raw] A pointer to the pixbuf's pixel data.  Please see 
for information about how the pixel data is stored in
memory.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
