\alias{gdkPixbufGetRowstride}
\name{gdkPixbufGetRowstride}
\title{gdkPixbufGetRowstride}
\description{Queries the rowstride of a pixbuf, which is the number of bytes between the start of a row
and the start of the next row.}
\usage{gdkPixbufGetRowstride(object)}
\arguments{\item{\verb{object}}{A pixbuf.}}
\value{[integer] Distance between row starts.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
