\alias{gdkPixbufLoaderNew}
\name{gdkPixbufLoaderNew}
\title{gdkPixbufLoaderNew}
\description{Creates a new pixbuf loader object.}
\usage{gdkPixbufLoaderNew()}
\value{[\verb{\link{GdkPixbufLoader}}]  A newly-created pixbuf loader.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
