\alias{gdkPixbufRotateSimple}
\name{gdkPixbufRotateSimple}
\title{gdkPixbufRotateSimple}
\description{Rotates a pixbuf by a multiple of 90 degrees, and returns the
result in a new pixbuf.}
\usage{gdkPixbufRotateSimple(object, angle)}
\arguments{
\item{\code{object}}{[\code{\link{GdkPixbuf}}]  a \code{\link{GdkPixbuf}}}
\item{\code{angle}}{[\code{\link{GdkPixbufRotation}}]  the angle to rotate by}
}
\details{  Since  2.6}
\value{[\code{\link{GdkPixbuf}}]  a new pixbuf}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
