\alias{gdkPixbufSimpleAnimGetLoop}
\name{gdkPixbufSimpleAnimGetLoop}
\title{gdkPixbufSimpleAnimGetLoop}
\description{Gets whether \code{animation} should loop indefinitely when it reaches the end.}
\usage{gdkPixbufSimpleAnimGetLoop(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkPixbufSimpleAnim}}}}
\details{Since 2.18}
\value{[logical] \code{TRUE} if the animation loops forever, \code{FALSE} otherwise}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
