\alias{gdkPixbufSimpleAnimNew}
\name{gdkPixbufSimpleAnimNew}
\title{gdkPixbufSimpleAnimNew}
\description{Creates a new, empty animation.}
\usage{gdkPixbufSimpleAnimNew(width, height, rate)}
\arguments{
\item{\verb{width}}{[integer]  the width of the animation}
\item{\verb{height}}{[integer]  the height of the animation}
\item{\verb{rate}}{[numeric]  the speed of the animation, in frames per second}
}
\details{  Since  2.8}
\value{[\verb{\link{GdkPixbufSimpleAnim}}]  a newly allocated \verb{\link{GdkPixbufSimpleAnim}} }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
