\alias{gdkPixbufSimpleAnimSetLoop}
\name{gdkPixbufSimpleAnimSetLoop}
\title{gdkPixbufSimpleAnimSetLoop}
\description{Sets whether \code{animation} should loop indefinitely when it reaches the end.}
\usage{gdkPixbufSimpleAnimSetLoop(object, loop)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkPixbufSimpleAnim}}}
\item{\verb{loop}}{whether to loop the animation}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
