\alias{gdkPointerIsGrabbed}
\name{gdkPointerIsGrabbed}
\title{gdkPointerIsGrabbed}
\description{Returns \code{TRUE} if the pointer on the default display is currently 
grabbed by this application.}
\usage{gdkPointerIsGrabbed()}
\details{Note that this does not take the inmplicit pointer grab on button
presses into account.
  
  }
\value{[logical] \code{TRUE} if the pointer is currently grabbed by this application.*}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
