\alias{gdkPropertyDelete}
\name{gdkPropertyDelete}
\title{gdkPropertyDelete}
\description{Deletes a property from a window.}
\usage{gdkPropertyDelete(object, property)}
\arguments{
\item{\code{object}}{[\code{\link{GdkWindow}}] a \code{\link{GdkWindow}}.}
\item{\code{property}}{[\code{\link{GdkAtom}}] the property to delete.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
