\alias{gdkRegionCopy}
\name{gdkRegionCopy}
\title{gdkRegionCopy}
\description{Copies \code{region}, creating an identical new region.}
\usage{gdkRegionCopy(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkRegion}}]  a \code{\link{GdkRegion}}}}
\value{[\code{\link{GdkRegion}}]  a new region identical to \code{region}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
