\alias{gdkRegionEmpty}
\name{gdkRegionEmpty}
\title{gdkRegionEmpty}
\description{Finds out if the \verb{\link{GdkRegion}} is empty.}
\usage{gdkRegionEmpty(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkRegion}}]  a \verb{\link{GdkRegion}}}}
\value{[logical] \code{TRUE} if \code{region} is empty.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
