\alias{gdkRegionEqual}
\name{gdkRegionEqual}
\title{gdkRegionEqual}
\description{Finds out if the two regions are the same.}
\usage{gdkRegionEqual(object, region2)}
\arguments{
\item{\code{object}}{[\code{\link{GdkRegion}}]  a \code{\link{GdkRegion}}}
\item{\code{region2}}{[\code{\link{GdkRegion}}]  a \code{\link{GdkRegion}}}
}
\value{[logical]  \code{TRUE} if \code{region1} and \code{region2} are equal.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
