\alias{gdkRegionGetClipbox}
\name{gdkRegionGetClipbox}
\title{gdkRegionGetClipbox}
\description{Obtains the smallest rectangle which includes the entire \verb{\link{GdkRegion}}.}
\usage{gdkRegionGetClipbox(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkRegion}}]  a \verb{\link{GdkRegion}}}}
\value{
A list containing the following elements:
\item{\verb{rectangle}}{[\verb{\link{GdkRectangle}}]  return location for the clipbox}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
