\alias{gdkRegionIntersect}
\name{gdkRegionIntersect}
\title{gdkRegionIntersect}
\description{Sets the area of \code{source1} to the intersection of the areas of \code{source1}
and \code{source2}. The resulting area is the set of pixels contained in
both \code{source1} and \code{source2}.}
\usage{gdkRegionIntersect(object, source2)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkRegion}}}
\item{\verb{source2}}{another \code{\link{GdkRegion}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
