\alias{gdkRegionNew}
\name{gdkRegionNew}
\title{gdkRegionNew}
\description{Creates a new empty \code{\link{GdkRegion}}.}
\usage{gdkRegionNew()}
\value{[\code{\link{GdkRegion}}]  a new empty \code{\link{GdkRegion}}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
