\alias{gdkRegionOffset}
\name{gdkRegionOffset}
\title{gdkRegionOffset}
\description{Moves a region the specified distance.}
\usage{gdkRegionOffset(object, dx, dy)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkRegion}}]  a \verb{\link{GdkRegion}}}
\item{\verb{dx}}{[integer]  the distance to move the region horizontally}
\item{\verb{dy}}{[integer]  the distance to move the region vertically}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
