\alias{gdkRegionPointIn}
\name{gdkRegionPointIn}
\title{gdkRegionPointIn}
\description{Returns \code{TRUE} if a point is in a region.}
\usage{gdkRegionPointIn(object, x, y)}
\arguments{
\item{\code{object}}{[\code{\link{GdkRegion}}] a \code{\link{GdkRegion}}.}
\item{\code{x}}{[integer] the x coordinate of a point.}
\item{\code{y}}{[integer] the y coordinate of a point.}
}
\value{[logical] \code{TRUE} if the point is in \code{region}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
