\alias{gdkRegionRectEqual}
\name{gdkRegionRectEqual}
\title{gdkRegionRectEqual}
\description{Finds out if a regions is the same as a rectangle.}
\usage{gdkRegionRectEqual(object, rectangle)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkRegion}}}
\item{\verb{rectangle}}{a \code{\link{GdkRectangle}}}
}
\details{Since 2.18}
\value{[logical] \code{TRUE} if \code{region} and \code{rectangle} are equal.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
