\alias{gdkRegionShrink}
\name{gdkRegionShrink}
\title{gdkRegionShrink}
\description{Resizes a region by the specified amount.
Positive values shrink the region. Negative values expand it.}
\usage{gdkRegionShrink(object, dx, dy)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkRegion}}]  a \verb{\link{GdkRegion}}}
\item{\verb{dx}}{[integer]  the number of pixels to shrink the region horizontally}
\item{\verb{dy}}{[integer]  the number of pixels to shrink the region vertically}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
