\alias{gdkRegionSubtract}
\name{gdkRegionSubtract}
\title{gdkRegionSubtract}
\description{Subtracts the area of \code{source2} from the area \code{source1}. The resulting
area is the set of pixels contained in \code{source1} but not in \code{source2}.}
\usage{gdkRegionSubtract(object, source2)}
\arguments{
\item{\code{object}}{[\code{\link{GdkRegion}}]  a \code{\link{GdkRegion}}}
\item{\code{source2}}{[\code{\link{GdkRegion}}]  another \code{\link{GdkRegion}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
