\alias{gdkRegionUnion}
\name{gdkRegionUnion}
\title{gdkRegionUnion}
\description{Sets the area of \code{source1} to the union of the areas of \code{source1} and
\code{source2}. The resulting area is the set of pixels contained in
either \code{source1} or \code{source2}.}
\usage{gdkRegionUnion(object, source2)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkRegion}}}
\item{\verb{source2}}{a \code{\link{GdkRegion}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
