\alias{gdkRgbDitherable}
\name{gdkRgbDitherable}
\title{gdkRgbDitherable}
\description{Determines whether the preferred visual is ditherable. This function may be
useful for presenting a user interface choice to the user about which
dither mode is desired; if the display is not ditherable, it may make
sense to gray out or hide the corresponding UI widget.}
\usage{gdkRgbDitherable()}
\value{[logical] \code{TRUE} if the preferred visual is ditherable.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
