\alias{gdkRgbSetVerbose}
\name{gdkRgbSetVerbose}
\title{gdkRgbSetVerbose}
\description{Sets the "verbose" flag. This is generally only useful for debugging.}
\usage{gdkRgbSetVerbose(verbose)}
\arguments{\item{\verb{verbose}}{\code{TRUE} if verbose messages are desired.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
