\alias{gdkScreenGetDefault}
\name{gdkScreenGetDefault}
\title{gdkScreenGetDefault}
\description{Gets the default screen for the default display. (See
\code{\link{gdkDisplayGetDefault}}).}
\usage{gdkScreenGetDefault()}
\details{  Since  2.2}
\value{[\verb{\link{GdkScreen}}]  a \verb{\link{GdkScreen}}, or \code{NULL} if there is no default display.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
