\alias{gdkScreenGetDefaultColormap}
\name{gdkScreenGetDefaultColormap}
\title{gdkScreenGetDefaultColormap}
\description{Gets the default colormap for \code{screen}.}
\usage{gdkScreenGetDefaultColormap(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkScreen}}]  a \code{\link{GdkScreen}}}}
\details{  Since  2.2}
\value{[\code{\link{GdkColormap}}]  the default \code{\link{GdkColormap}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
