\alias{gdkScreenGetDisplay}
\name{gdkScreenGetDisplay}
\title{gdkScreenGetDisplay}
\description{Gets the display to which the \code{screen} belongs.}
\usage{gdkScreenGetDisplay(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkScreen}}]  a \verb{\link{GdkScreen}}}}
\details{  Since  2.2}
\value{[\verb{\link{GdkDisplay}}]  the display to which \code{screen} belongs}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
