\alias{gdkScreenGetHeight}
\name{gdkScreenGetHeight}
\title{gdkScreenGetHeight}
\description{Gets the height of \code{screen} in pixels}
\usage{gdkScreenGetHeight(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkScreen}}}}
\details{Since 2.2}
\value{[integer] the height of \code{screen} in pixels.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
