\alias{gdkScreenGetHeightMm}
\name{gdkScreenGetHeightMm}
\title{gdkScreenGetHeightMm}
\description{Returns the height of \code{screen} in millimeters. 
Note that on some X servers this value will not be correct.}
\usage{gdkScreenGetHeightMm(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkScreen}}}}
\details{Since 2.2}
\value{[integer] the heigth of \code{screen} in millimeters.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
