\alias{gdkScreenGetNMonitors}
\name{gdkScreenGetNMonitors}
\title{gdkScreenGetNMonitors}
\description{Returns the number of monitors which \code{screen} consists of.}
\usage{gdkScreenGetNMonitors(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkScreen}}}}
\details{Since 2.2}
\value{[integer] number of monitors which \code{screen} consists of}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
