\alias{gdkScreenGetNumber}
\name{gdkScreenGetNumber}
\title{gdkScreenGetNumber}
\description{Gets the index of \code{screen} among the screens in the display
to which it belongs. (See \code{\link{gdkScreenGetDisplay}})}
\usage{gdkScreenGetNumber(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkScreen}}}}
\details{Since 2.2}
\value{[integer] the index}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
