\alias{gdkScreenGetResolution}
\name{gdkScreenGetResolution}
\title{gdkScreenGetResolution}
\description{Gets the resolution for font handling on the screen; see
\code{\link{gdkScreenSetResolution}} for full details.}
\usage{gdkScreenGetResolution(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkScreen}}]  a \code{\link{GdkScreen}}}}
\details{  Since  2.10}
\value{[numeric]  the current resolution, or -1 if no resolution
has been set.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
