\alias{gdkScreenGetRootWindow}
\name{gdkScreenGetRootWindow}
\title{gdkScreenGetRootWindow}
\description{Gets the root window of \code{screen}.}
\usage{gdkScreenGetRootWindow(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkScreen}}]  a \verb{\link{GdkScreen}}}}
\details{  Since  2.2}
\value{[\verb{\link{GdkWindow}}]  the root window}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
