\alias{gdkScreenGetSystemVisual}
\name{gdkScreenGetSystemVisual}
\title{gdkScreenGetSystemVisual}
\description{Get the system's default visual for \code{screen}.
This is the visual for the root window of the display.}
\usage{gdkScreenGetSystemVisual(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkScreen}}]  a \code{\link{GdkScreen}}.}}
\details{  Since  2.2}
\value{[\code{\link{GdkVisual}}]  the system visual}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
