\alias{gdkScreenGetWidthMm}
\name{gdkScreenGetWidthMm}
\title{gdkScreenGetWidthMm}
\description{Gets the width of \code{screen} in millimeters. 
Note that on some X servers this value will not be correct.}
\usage{gdkScreenGetWidthMm(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkScreen}}}}
\details{Since 2.2}
\value{[integer] the width of \code{screen} in millimeters.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
