\alias{gdkScreenHeight}
\name{gdkScreenHeight}
\title{gdkScreenHeight}
\description{Returns the height of the default screen in pixels.}
\usage{gdkScreenHeight()}
\value{[integer]  the height of the default screen in pixels.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
