\alias{gdkScreenHeightMm}
\name{gdkScreenHeightMm}
\title{gdkScreenHeightMm}
\description{Returns the height of the default screen in millimeters.
Note that on many X servers this value will not be correct.}
\usage{gdkScreenHeightMm()}
\value{[integer] the height of the default screen in millimeters,
though it is not always correct.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
