\alias{gdkScreenSetDefaultColormap}
\name{gdkScreenSetDefaultColormap}
\title{gdkScreenSetDefaultColormap}
\description{Sets the default \code{colormap} for \code{screen}.}
\usage{gdkScreenSetDefaultColormap(object, colormap)}
\arguments{
\item{\code{object}}{[\code{\link{GdkScreen}}]  a \code{\link{GdkScreen}}}
\item{\code{colormap}}{[\code{\link{GdkColormap}}]  a \code{\link{GdkColormap}}}
}
\details{  Since  2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
