\alias{gdkScreenWidth}
\name{gdkScreenWidth}
\title{gdkScreenWidth}
\description{Returns the width of the default screen in pixels.}
\usage{gdkScreenWidth()}
\value{[integer]  the width of the default screen in pixels.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
