\alias{gdkScreenWidthMm}
\name{gdkScreenWidthMm}
\title{gdkScreenWidthMm}
\description{Returns the width of the default screen in millimeters.
Note that on many X servers this value will not be correct.}
\usage{gdkScreenWidthMm()}
\value{[integer] the width of the default screen in millimeters,
though it is not always correct.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
