\alias{gdkSetDoubleClickTime}
\name{gdkSetDoubleClickTime}
\title{gdkSetDoubleClickTime}
\description{Set the double click time for the default display. See
\code{\link{gdkDisplaySetDoubleClickTime}}. 
See also \code{\link{gdkDisplaySetDoubleClickDistance}}.
Applications should \emph{not} set this, it is a 
global user-configured setting.}
\usage{gdkSetDoubleClickTime(msec)}
\arguments{\item{\verb{msec}}{double click time in milliseconds (thousandths of a second)}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
