\alias{gdkSetProgramClass}
\name{gdkSetProgramClass}
\title{gdkSetProgramClass}
\description{Sets the program class. The X11 backend uses the program class to set
the class name part of the \code{WM_CLASS} property on
toplevel windows; see the ICCCM.}
\usage{gdkSetProgramClass(program.class)}
\arguments{\item{\verb{program.class}}{a string.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
