\alias{gdkSetShowEvents}
\name{gdkSetShowEvents}
\title{gdkSetShowEvents}
\description{Sets whether a trace of received events is output.
Note that GTK+ must be compiled with debugging (that is,
configured using the \option{--enable-debug} option)
to use this option.}
\usage{gdkSetShowEvents(show.events)}
\arguments{\item{\verb{show.events}}{\code{TRUE} to output event debugging information.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
