\alias{gdkVisualGetBestDepth}
\name{gdkVisualGetBestDepth}
\title{gdkVisualGetBestDepth}
\description{Get the best available depth for the default GDK screen.  "Best"
means "largest," i.e. 32 preferred over 24 preferred over 8 bits
per pixel.}
\usage{gdkVisualGetBestDepth()}
\value{[integer] best available depth}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
