\alias{gdkVisualGetBestType}
\name{gdkVisualGetBestType}
\title{gdkVisualGetBestType}
\description{Return the best available visual type for the default GDK screen.}
\usage{gdkVisualGetBestType()}
\value{[\code{\link{GdkVisualType}}]  best visual type}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
