\alias{gdkVisualGetScreen}
\name{gdkVisualGetScreen}
\title{gdkVisualGetScreen}
\description{Gets the screen to which this visual belongs}
\usage{gdkVisualGetScreen(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkVisual}}]  a \verb{\link{GdkVisual}}}}
\details{  Since  2.2}
\value{[\verb{\link{GdkScreen}}]  the screen to which this visual belongs.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
