\alias{gdkVisualGetSystem}
\name{gdkVisualGetSystem}
\title{gdkVisualGetSystem}
\description{Get the system'sdefault visual for the default GDK screen.
This is the visual for the root window of the display.}
\usage{gdkVisualGetSystem()}
\value{[\verb{\link{GdkVisual}}]  system visual}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
