\alias{gdkWindowBeep}
\name{gdkWindowBeep}
\title{gdkWindowBeep}
\description{Emits a short beep associated to \code{window} in the appropriate
display, if supported. Otherwise, emits a short beep on
the display just as \code{\link{gdkDisplayBeep}}.}
\usage{gdkWindowBeep(object)}
\arguments{\item{\verb{object}}{a toplevel \code{\link{GdkWindow}}}}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
