\alias{gdkWindowClearArea}
\name{gdkWindowClearArea}
\title{gdkWindowClearArea}
\description{Clears an area of \code{window} to the background color or background pixmap.}
\usage{gdkWindowClearArea(object, x, y, width, height)}
\arguments{
\item{\code{object}}{[\code{\link{GdkWindow}}]  a \code{\link{GdkWindow}}}
\item{\code{x}}{[integer]  x coordinate of rectangle to clear}
\item{\code{y}}{[integer]  y coordinate of rectangle to clear}
\item{\code{width}}{[integer]  width of rectangle to clear}
\item{\code{height}}{[integer]  height of rectangle to clear}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
