\alias{gdkWindowGeometryChanged}
\name{gdkWindowGeometryChanged}
\title{gdkWindowGeometryChanged}
\description{This function informs GDK that the geometry of an embedded
offscreen window has changed. This is necessary for GDK to keep
track of which offscreen window the pointer is in.}
\usage{gdkWindowGeometryChanged(object)}
\arguments{\item{\verb{object}}{an embedded offscreen \code{\link{GdkWindow}}}}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
