\alias{gdkWindowGetEvents}
\name{gdkWindowGetEvents}
\title{gdkWindowGetEvents}
\description{Gets the event mask for \code{window}. See \code{\link{gdkWindowSetEvents}}.}
\usage{gdkWindowGetEvents(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkWindow}}]  a \code{\link{GdkWindow}}}}
\value{[\code{\link{GdkEventMask}}]  event mask for \code{window}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
