\alias{gdkWindowGetGroup}
\name{gdkWindowGetGroup}
\title{gdkWindowGetGroup}
\description{Returns the group leader window for \code{window}. See \code{\link{gdkWindowSetGroup}}.}
\usage{gdkWindowGetGroup(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkWindow}}]  a toplevel \code{\link{GdkWindow}}}}
\details{  Since  2.4}
\value{[\code{\link{GdkWindow}}]  the group leader window for \code{window}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
