\alias{gdkWindowGetToplevel}
\name{gdkWindowGetToplevel}
\title{gdkWindowGetToplevel}
\description{Gets the toplevel window that's an ancestor of \code{window}.}
\usage{gdkWindowGetToplevel(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkWindow}}]  a \code{\link{GdkWindow}}}}
\value{[\code{\link{GdkWindow}}]  the toplevel window containing \code{window}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
