\alias{gdkWindowGetToplevels}
\name{gdkWindowGetToplevels}
\title{gdkWindowGetToplevels}
\description{Obtains a list of all toplevel windows known to GDK on the default
screen (see \code{gdkWindowGetToplevelsForScreen()}).
A toplevel window is a child of the root window (see
\code{\link{gdkGetDefaultRootWindow}}).}
\usage{gdkWindowGetToplevels()}
\details{  .}
\value{[list] .}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
