\alias{gdkWindowGetTypeHint}
\name{gdkWindowGetTypeHint}
\title{gdkWindowGetTypeHint}
\description{This function returns the type hint set for a window.}
\usage{gdkWindowGetTypeHint(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkWindow}}]  A toplevel \code{\link{GdkWindow}}}}
\details{  Since  2.10}
\value{[\code{\link{GdkWindowTypeHint}}]  The type hint set for \code{window}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
