\alias{gdkWindowGetWindowType}
\name{gdkWindowGetWindowType}
\title{gdkWindowGetWindowType}
\description{Gets the type of the window. See \code{\link{GdkWindowType}}.}
\usage{gdkWindowGetWindowType(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkWindow}}]  a \code{\link{GdkWindow}}}}
\value{[\code{\link{GdkWindowType}}]  type of window}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
