\alias{gdkWindowHide}
\name{gdkWindowHide}
\title{gdkWindowHide}
\description{For toplevel windows, withdraws them, so they will no longer be
known to the window manager; for all windows, unmaps them, so
they won't be displayed. Normally done automatically as
part of \code{\link{gtkWidgetHide}}.}
\usage{gdkWindowHide(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkWindow}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
