\alias{gdkWindowIsDestroyed}
\name{gdkWindowIsDestroyed}
\title{gdkWindowIsDestroyed}
\description{Check to see if a window is destroyed..}
\usage{gdkWindowIsDestroyed(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkWindow}}}}
\details{Since 2.18}
\value{[logical] \code{TRUE} if the window is destroyed}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
