\alias{gdkWindowIsVisible}
\name{gdkWindowIsVisible}
\title{gdkWindowIsVisible}
\description{Checks whether the window has been mapped (with \code{\link{gdkWindowShow}} or
\code{\link{gdkWindowShowUnraised}}).}
\usage{gdkWindowIsVisible(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkWindow}}]  a \code{\link{GdkWindow}}}}
\value{[logical]  \code{TRUE} if the window is mapped}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
