\alias{gdkWindowRemoveRedirection}
\name{gdkWindowRemoveRedirection}
\title{gdkWindowRemoveRedirection}
\description{Removes any active redirection started by
\code{\link{gdkWindowRedirectToDrawable}}.}
\usage{gdkWindowRemoveRedirection(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkWindow}}}}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
