\alias{gdkWindowSetStartupId}
\name{gdkWindowSetStartupId}
\title{gdkWindowSetStartupId}
\description{When using GTK+, typically you should use \code{\link{gtkWindowSetStartupId}}
instead of this low-level function.}
\usage{gdkWindowSetStartupId(object, startup.id)}
\arguments{
\item{\code{object}}{[\code{\link{GdkWindow}}]  a toplevel \code{\link{GdkWindow}}}
\item{\code{startup.id}}{[character]  a string with startup-notification identifier}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
