\alias{gdkWindowUnstick}
\name{gdkWindowUnstick}
\title{gdkWindowUnstick}
\description{Reverse operation for \code{\link{gdkWindowStick}}; see \code{\link{gdkWindowStick}},
and \code{\link{gtkWindowUnstick}}.}
\usage{gdkWindowUnstick(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkWindow}}]  a toplevel \verb{\link{GdkWindow}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
