\alias{gDriveCanEject}
\name{gDriveCanEject}
\title{gDriveCanEject}
\description{Checks if a drive can be ejected.}
\usage{gDriveCanEject(object)}
\arguments{\item{\verb{object}}{a \code{\link{GDrive}}.}}
\value{[logical] \code{TRUE} if the \code{drive} can be ejected, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
