\alias{gDriveCanPollForMedia}
\name{gDriveCanPollForMedia}
\title{gDriveCanPollForMedia}
\description{Checks if a drive can be polled for media changes.}
\usage{gDriveCanPollForMedia(object)}
\arguments{\item{\verb{object}}{a \code{\link{GDrive}}.}}
\value{[logical] \code{TRUE} if the \code{drive} can be polled for media changes,
\code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
