\alias{gDriveCanStartDegraded}
\name{gDriveCanStartDegraded}
\title{gDriveCanStartDegraded}
\description{Checks if a drive can be started degraded.}
\usage{gDriveCanStartDegraded(object)}
\arguments{\item{\verb{object}}{a \code{\link{GDrive}}.}}
\details{Since 2.22}
\value{[logical] \code{TRUE} if the \code{drive} can be started degraded, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
