\alias{gDriveCanStop}
\name{gDriveCanStop}
\title{gDriveCanStop}
\description{Checks if a drive can be stopped.}
\usage{gDriveCanStop(object)}
\arguments{\item{\verb{object}}{a \code{\link{GDrive}}.}}
\details{Since 2.22}
\value{[logical] \code{TRUE} if the \code{drive} can be stopped, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
