\alias{gDriveEnumerateIdentifiers}
\name{gDriveEnumerateIdentifiers}
\title{gDriveEnumerateIdentifiers}
\description{Gets the kinds of identifiers that \code{drive} has. 
Use \code{gDriveGetIdentifer()} to obtain the identifiers
themselves.}
\usage{gDriveEnumerateIdentifiers(object)}
\arguments{\item{\verb{object}}{a \code{\link{GDrive}}}}
\value{[char] a list of strings containing
kinds of identifiers.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
