\alias{gDriveGetIcon}
\name{gDriveGetIcon}
\title{gDriveGetIcon}
\description{Gets the icon for \code{drive}.}
\usage{gDriveGetIcon(object)}
\arguments{\item{\verb{object}}{a \code{\link{GDrive}}.}}
\value{[\code{\link{GIcon}}] \code{\link{GIcon}} for the \code{drive}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
