\alias{gDriveGetName}
\name{gDriveGetName}
\title{gDriveGetName}
\description{Gets the name of \code{drive}.}
\usage{gDriveGetName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GDrive}}.}}
\value{[char] a string containing \code{drive}'s name.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
