\alias{gDriveGetStartStopType}
\name{gDriveGetStartStopType}
\title{gDriveGetStartStopType}
\description{Gets a hint about how a drive can be started/stopped.}
\usage{gDriveGetStartStopType(object)}
\arguments{\item{\verb{object}}{a \code{\link{GDrive}}.}}
\details{Since 2.22}
\value{[\code{\link{GDriveStartStopType}}] A value from the \code{\link{GDriveStartStopType}} enumeration.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
