\alias{gDriveGetVolumes}
\name{gDriveGetVolumes}
\title{gDriveGetVolumes}
\description{Get a list of mountable volumes for \code{drive}.}
\usage{gDriveGetVolumes(object)}
\arguments{\item{\verb{object}}{a \code{\link{GDrive}}.}}
\value{[list] \verb{list} containing any \code{\link{GVolume}} objects on the given \code{drive}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
