\alias{gDriveHasVolumes}
\name{gDriveHasVolumes}
\title{gDriveHasVolumes}
\description{Check if \code{drive} has any mountable volumes.}
\usage{gDriveHasVolumes(object)}
\arguments{\item{\verb{object}}{a \code{\link{GDrive}}.}}
\value{[logical] \code{TRUE} if the \code{drive} contains volumes, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
