\alias{gDriveIsMediaCheckAutomatic}
\name{gDriveIsMediaCheckAutomatic}
\title{gDriveIsMediaCheckAutomatic}
\description{Checks if \code{drive} is capabable of automatically detecting media changes.}
\usage{gDriveIsMediaCheckAutomatic(object)}
\arguments{\item{\verb{object}}{a \code{\link{GDrive}}.}}
\value{[logical] \code{TRUE} if the \code{drive} is capabable of automatically detecting
media changes, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
