\alias{gDriveIsMediaRemovable}
\name{gDriveIsMediaRemovable}
\title{gDriveIsMediaRemovable}
\description{Checks if the \code{drive} supports removable media.}
\usage{gDriveIsMediaRemovable(object)}
\arguments{\item{\verb{object}}{a \code{\link{GDrive}}.}}
\value{[logical] \code{TRUE} if \code{drive} supports removable media, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
