\alias{gEmblemedIconAddEmblem}
\name{gEmblemedIconAddEmblem}
\title{gEmblemedIconAddEmblem}
\description{Adds \code{emblem} to the \verb{list} of \code{\link{GEmblem}} s.}
\usage{gEmblemedIconAddEmblem(object, emblem)}
\arguments{
\item{\verb{object}}{a \code{\link{GEmblemedIcon}}}
\item{\verb{emblem}}{a \code{\link{GEmblem}}}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
