\alias{gEmblemedIconGetEmblems}
\name{gEmblemedIconGetEmblems}
\title{gEmblemedIconGetEmblems}
\description{Gets the list of emblems for the \code{icon}.}
\usage{gEmblemedIconGetEmblems(object)}
\arguments{\item{\verb{object}}{a \code{\link{GEmblemedIcon}}}}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
