\alias{gEmblemedIconGetIcon}
\name{gEmblemedIconGetIcon}
\title{gEmblemedIconGetIcon}
\description{Gets the main icon for \code{emblemed}.}
\usage{gEmblemedIconGetIcon(object)}
\arguments{\item{\verb{object}}{a \code{\link{GEmblemedIcon}}}}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
