\alias{gEmblemedIconNew}
\name{gEmblemedIconNew}
\title{gEmblemedIconNew}
\description{Creates a new emblemed icon for \code{icon} with the emblem \code{emblem}.}
\usage{gEmblemedIconNew(icon, emblem)}
\arguments{
\item{\verb{icon}}{a \code{\link{GIcon}}}
\item{\verb{emblem}}{a \code{\link{GEmblem}}}
}
\details{Since 2.18}
\value{[\code{\link{GIcon}}] a new \code{\link{GIcon}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
