\alias{gEmblemGetIcon}
\name{gEmblemGetIcon}
\title{gEmblemGetIcon}
\description{Gives back the icon from \code{emblem}.}
\usage{gEmblemGetIcon(object)}
\arguments{\item{\verb{object}}{a \code{\link{GEmblem}} from which the icon should be extracted.}}
\details{Since 2.18}
\value{[\code{\link{GIcon}}] a \code{\link{GIcon}}. The returned object belongs to the emblem
and should not be modified or freed.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
