\alias{gEmblemGetOrigin}
\name{gEmblemGetOrigin}
\title{gEmblemGetOrigin}
\description{Gets the origin of the emblem.}
\usage{gEmblemGetOrigin(object)}
\arguments{\item{\verb{object}}{a \code{\link{GEmblem}}}}
\details{Since 2.18}
\value{[\code{\link{GEmblemOrigin}}] the origin of the emblem}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
