\alias{gEmblemNew}
\name{gEmblemNew}
\title{gEmblemNew}
\description{Creates a new emblem for \code{icon}.}
\usage{gEmblemNew(icon = NULL, origin = NULL)}
\arguments{
\item{\verb{icon}}{a GIcon containing the icon.}
\item{\verb{origin}}{a new \code{\link{GEmblem}}.}
}
\details{Since 2.18}
\value{[\code{\link{GEmblem}}] a new \code{\link{GEmblem}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
