\alias{gEmblemNewWithOrigin}
\name{gEmblemNewWithOrigin}
\title{gEmblemNewWithOrigin}
\description{Creates a new emblem for \code{icon}.}
\usage{gEmblemNewWithOrigin(icon, origin)}
\arguments{
\item{\verb{icon}}{a GIcon containing the icon.}
\item{\verb{origin}}{a GEmblemOrigin enum defining the emblem's origin}
}
\details{Since 2.18}
\value{[\code{\link{GEmblem}}] a new \code{\link{GEmblem}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
