\alias{atkKeyEventStructGetKeycode}
\alias{atkKeyEventStructGetKeyval}
\alias{atkKeyEventStructGetLength}
\alias{atkKeyEventStructGetState}
\alias{atkKeyEventStructGetString}
\alias{atkKeyEventStructGetTimestamp}
\alias{cairoFontExtentsGetAscent}
\alias{cairoFontExtentsGetDescent}
\alias{cairoFontExtentsGetHeight}
\alias{cairoFontExtentsGetMaxXAdvance}
\alias{cairoFontExtentsGetMaxYAdvance}
\alias{cairoMatrixGetX0}
\alias{cairoMatrixGetXx}
\alias{cairoMatrixGetXy}
\alias{cairoMatrixGetY0}
\alias{cairoMatrixGetYx}
\alias{cairoMatrixGetYy}
\alias{cairoTextExtentsGetHeight}
\alias{cairoTextExtentsGetWidth}
\alias{cairoTextExtentsGetXAdvance}
\alias{cairoTextExtentsGetXBearing}
\alias{cairoTextExtentsGetYAdvance}
\alias{cairoTextExtentsGetYBearing}
\alias{gdkDeviceAxisGetMax}
\alias{gdkDeviceAxisGetMin}
\alias{gdkDeviceAxisGetUse}
\alias{gdkDeviceGetAxes}
\alias{gdkDeviceGetHasCursor}
\alias{gdkDeviceGetKeys}
\alias{gdkDeviceGetMode}
\alias{gdkDeviceGetName}
\alias{gdkDeviceGetNumAxes}
\alias{gdkDeviceGetNumKeys}
\alias{gdkDeviceGetSource}
\alias{gdkDeviceKeyGetKeyval}
\alias{gdkDeviceKeyGetModifiers}
\alias{gdkDragContextGetAction}
\alias{gdkDragContextGetActions}
\alias{gdkDragContextGetDestWindow}
\alias{gdkDragContextGetIsSource}
\alias{gdkDragContextGetProtocol}
\alias{gdkDragContextGetSourceWindow}
\alias{gdkDragContextGetStartTime}
\alias{gdkDragContextGetSuggestedAction}
\alias{gdkDragContextGetTargets}
\alias{gdkEventAnyGetSendEvent}
\alias{gdkEventAnyGetWindow}
\alias{gdkEventButtonGetAxes}
\alias{gdkEventButtonGetButton}
\alias{gdkEventButtonGetDevice}
\alias{gdkEventButtonGetState}
\alias{gdkEventButtonGetTime}
\alias{gdkEventButtonGetX}
\alias{gdkEventButtonGetXRoot}
\alias{gdkEventButtonGetY}
\alias{gdkEventButtonGetYRoot}
\alias{gdkEventClientGetMessageType}
\alias{gdkEventConfigureGetHeight}
\alias{gdkEventConfigureGetWidth}
\alias{gdkEventConfigureGetX}
\alias{gdkEventConfigureGetY}
\alias{gdkEventCrossingGetDetail}
\alias{gdkEventCrossingGetFocus}
\alias{gdkEventCrossingGetMode}
\alias{gdkEventCrossingGetState}
\alias{gdkEventCrossingGetTime}
\alias{gdkEventCrossingGetX}
\alias{gdkEventCrossingGetXRoot}
\alias{gdkEventCrossingGetY}
\alias{gdkEventCrossingGetYRoot}
\alias{gdkEventDNDGetContext}
\alias{gdkEventDNDGetTime}
\alias{gdkEventDNDGetXRoot}
\alias{gdkEventDNDGetYRoot}
\alias{gdkEventExposeGetArea}
\alias{gdkEventExposeGetCount}
\alias{gdkEventExposeGetRegion}
\alias{gdkEventFocusGetIn}
\alias{gdkEventGrabBrokenGetGrabWindow}
\alias{gdkEventGrabBrokenGetImplicit}
\alias{gdkEventGrabBrokenGetKeyboard}
\alias{gdkEventKeyGetGroup}
\alias{gdkEventKeyGetHardwareKeycode}
\alias{gdkEventKeyGetKeyval}
\alias{gdkEventKeyGetLength}
\alias{gdkEventKeyGetState}
\alias{gdkEventKeyGetString}
\alias{gdkEventKeyGetTime}
\alias{gdkEventMotionGetAxes}
\alias{gdkEventMotionGetDevice}
\alias{gdkEventMotionGetIsHint}
\alias{gdkEventMotionGetState}
\alias{gdkEventMotionGetTime}
\alias{gdkEventMotionGetX}
\alias{gdkEventMotionGetXRoot}
\alias{gdkEventMotionGetY}
\alias{gdkEventMotionGetYRoot}
\alias{gdkEventOwnerChangeGetOwner}
\alias{gdkEventOwnerChangeGetReason}
\alias{gdkEventOwnerChangeGetSelection}
\alias{gdkEventOwnerChangeGetSelectionTime}
\alias{gdkEventOwnerChangeGetTime}
\alias{gdkEventPropertyGetAtom}
\alias{gdkEventPropertyGetState}
\alias{gdkEventPropertyGetTime}
\alias{gdkEventProximityGetDevice}
\alias{gdkEventProximityGetTime}
\alias{gdkEventScrollGetDevice}
\alias{gdkEventScrollGetDirection}
\alias{gdkEventScrollGetState}
\alias{gdkEventScrollGetTime}
\alias{gdkEventScrollGetX}
\alias{gdkEventScrollGetXRoot}
\alias{gdkEventScrollGetY}
\alias{gdkEventScrollGetYRoot}
\alias{gdkEventSelectionGetProperty}
\alias{gdkEventSelectionGetRequestor}
\alias{gdkEventSelectionGetSelection}
\alias{gdkEventSelectionGetTarget}
\alias{gdkEventSelectionGetTime}
\alias{gdkEventSettingGetAction}
\alias{gdkEventSettingGetName}
\alias{gdkEventVisibilityGetState}
\alias{gdkEventWindowStateGetChangedMask}
\alias{gdkEventWindowStateGetNewWindowState}
\alias{gdkFontGetAscent}
\alias{gdkFontGetDescent}
\alias{gdkPangoAttrEmbossedGetEmbossed}
\alias{gdkPangoAttrStippleGetStipple}
\alias{gdkVisualGetBitsPerRgb}
\alias{gdkVisualGetBlueMask}
\alias{gdkVisualGetBluePrec}
\alias{gdkVisualGetBlueShift}
\alias{gdkVisualGetByteOrder}
\alias{gdkVisualGetColormapSize}
\alias{gdkVisualGetDepth}
\alias{gdkVisualGetGreenMask}
\alias{gdkVisualGetGreenPrec}
\alias{gdkVisualGetGreenShift}
\alias{gdkVisualGetRedMask}
\alias{gdkVisualGetRedPrec}
\alias{gdkVisualGetRedShift}
\alias{gtkAccelKeyGetAccelFlags}
\alias{gtkAccelKeyGetAccelKey}
\alias{gtkAccelKeyGetAccelMods}
\alias{gtkBoxChildGetExpand}
\alias{gtkBoxChildGetFill}
\alias{gtkBoxChildGetIsSecondary}
\alias{gtkBoxChildGetPack}
\alias{gtkBoxChildGetPadding}
\alias{gtkBoxChildGetWidget}
\alias{gtkCListRowGetBackground}
\alias{gtkCListRowGetBgSet}
\alias{gtkCListRowGetCell}
\alias{gtkCListRowGetData}
\alias{gtkCListRowGetDestroy}
\alias{gtkCListRowGetFgSet}
\alias{gtkCListRowGetForeground}
\alias{gtkCListRowGetSelectable}
\alias{gtkCListRowGetState}
\alias{gtkCListRowGetStyle}
\alias{gtkCTreeRowGetChildren}
\alias{gtkCTreeRowGetExpanded}
\alias{gtkCTreeRowGetIsLeaf}
\alias{gtkCTreeRowGetLevel}
\alias{gtkCTreeRowGetMaskClosed}
\alias{gtkCTreeRowGetMaskOpened}
\alias{gtkCTreeRowGetParent}
\alias{gtkCTreeRowGetPixmapClosed}
\alias{gtkCTreeRowGetPixmapOpened}
\alias{gtkCTreeRowGetRow}
\alias{gtkCTreeRowGetSibling}
\alias{gtkColorSelectionDialogGetCancelButton}
\alias{gtkColorSelectionDialogGetColorsel}
\alias{gtkColorSelectionDialogGetHelpButton}
\alias{gtkColorSelectionDialogGetOkButton}
\alias{gtkComboGetEntry}
\alias{gtkComboGetList}
\alias{gtkContainerGetHasFocusChain}
\alias{gtkContainerGetNeedResize}
\alias{gtkContainerGetReallocateRedraws}
\alias{gtkDialogGetVbox}
\alias{gtkFileSelectionGetActionArea}
\alias{gtkFileSelectionGetButtonArea}
\alias{gtkFileSelectionGetCancelButton}
\alias{gtkFileSelectionGetDirList}
\alias{gtkFileSelectionGetFileList}
\alias{gtkFileSelectionGetFileopCDir}
\alias{gtkFileSelectionGetFileopDelFile}
\alias{gtkFileSelectionGetFileopDialog}
\alias{gtkFileSelectionGetFileopEntry}
\alias{gtkFileSelectionGetFileopFile}
\alias{gtkFileSelectionGetFileopRenFile}
\alias{gtkFileSelectionGetHelpButton}
\alias{gtkFileSelectionGetHistoryMenu}
\alias{gtkFileSelectionGetHistoryPulldown}
\alias{gtkFileSelectionGetMainVbox}
\alias{gtkFileSelectionGetOkButton}
\alias{gtkFileSelectionGetSelectionEntry}
\alias{gtkFileSelectionGetSelectionText}
\alias{gtkFixedChildGetWidget}
\alias{gtkFixedChildGetX}
\alias{gtkFixedChildGetY}
\alias{gtkFixedGetChildren}
\alias{gtkGammaCurveGetCurve}
\alias{gtkGammaCurveGetGamma}
\alias{gtkGammaCurveGetGammaDialog}
\alias{gtkGammaCurveGetGammaText}
\alias{gtkGammaCurveGetTable}
\alias{gtkMessageDialogGetLabel}
\alias{gtkPreviewInfoGetGamma}
\alias{gtkPreviewInfoGetLookup}
\alias{gtkRequisitionGetHeight}
\alias{gtkRequisitionGetWidth}
\alias{gtkStyleGetBase}
\alias{gtkStyleGetBaseGc}
\alias{gtkStyleGetBg}
\alias{gtkStyleGetBgGc}
\alias{gtkStyleGetBgPixmap}
\alias{gtkStyleGetBlack}
\alias{gtkStyleGetBlackGc}
\alias{gtkStyleGetDark}
\alias{gtkStyleGetDarkGc}
\alias{gtkStyleGetFg}
\alias{gtkStyleGetFgGc}
\alias{gtkStyleGetFontDesc}
\alias{gtkStyleGetLight}
\alias{gtkStyleGetLightGc}
\alias{gtkStyleGetMid}
\alias{gtkStyleGetMidGc}
\alias{gtkStyleGetText}
\alias{gtkStyleGetTextAa}
\alias{gtkStyleGetTextAaGc}
\alias{gtkStyleGetTextGc}
\alias{gtkStyleGetWhite}
\alias{gtkStyleGetWhiteGc}
\alias{gtkStyleGetXthickness}
\alias{gtkStyleGetYthickness}
\alias{gtkTableChildGetBottomAttach}
\alias{gtkTableChildGetLeftAttach}
\alias{gtkTableChildGetRightAttach}
\alias{gtkTableChildGetTopAttach}
\alias{gtkTableChildGetWidget}
\alias{gtkTableChildGetXexpand}
\alias{gtkTableChildGetXfill}
\alias{gtkTableChildGetXpadding}
\alias{gtkTableChildGetXshrink}
\alias{gtkTableChildGetYexpand}
\alias{gtkTableChildGetYfill}
\alias{gtkTableChildGetYpadding}
\alias{gtkTableChildGetYshrink}
\alias{gtkTableGetChildren}
\alias{gtkTableGetCols}
\alias{gtkTableGetNcols}
\alias{gtkTableGetNrows}
\alias{gtkTableGetRows}
\alias{gtkTableRowColGetAllocation}
\alias{gtkTableRowColGetEmpty}
\alias{gtkTableRowColGetExpand}
\alias{gtkTableRowColGetNeedExpand}
\alias{gtkTableRowColGetNeedShrink}
\alias{gtkTableRowColGetRequisition}
\alias{gtkTableRowColGetShrink}
\alias{gtkTableRowColGetSpacing}
\alias{gtkTextAppearanceGetBgColor}
\alias{gtkTextAppearanceGetBgStipple}
\alias{gtkTextAppearanceGetDrawBg}
\alias{gtkTextAppearanceGetFgColor}
\alias{gtkTextAppearanceGetFgStipple}
\alias{gtkTextAppearanceGetRise}
\alias{gtkTextAppearanceGetStrikethrough}
\alias{gtkTextAppearanceGetUnderline}
\alias{gtkTextAttributesGetAppearance}
\alias{gtkTextAttributesGetBgFullHeight}
\alias{gtkTextAttributesGetDirection}
\alias{gtkTextAttributesGetEditable}
\alias{gtkTextAttributesGetFont}
\alias{gtkTextAttributesGetFontScale}
\alias{gtkTextAttributesGetIndent}
\alias{gtkTextAttributesGetInvisible}
\alias{gtkTextAttributesGetJustification}
\alias{gtkTextAttributesGetLanguage}
\alias{gtkTextAttributesGetLeftMargin}
\alias{gtkTextAttributesGetPixelsAboveLines}
\alias{gtkTextAttributesGetPixelsBelowLines}
\alias{gtkTextAttributesGetPixelsInsideWrap}
\alias{gtkTextAttributesGetRealized}
\alias{gtkTextAttributesGetRightMargin}
\alias{gtkTextAttributesGetTabs}
\alias{gtkTextAttributesGetWrapMode}
\alias{gtkToggleButtonGetDrawIndicator}
\alias{gtkWindowGetAllowGrow}
\alias{gtkWindowGetAllowShrink}
\alias{gtkWindowGetConfigureNotifyReceived}
\alias{gtkWindowGetConfigureRequestCount}
\alias{gtkWindowGetFocusWidget}
\alias{gtkWindowGetFrame}
\alias{gtkWindowGetFrameBottom}
\alias{gtkWindowGetFrameLeft}
\alias{gtkWindowGetFrameRight}
\alias{gtkWindowGetFrameTop}
\alias{gtkWindowGetHasFocus}
\alias{gtkWindowGetHasUserRefCount}
\alias{gtkWindowGetIconifyInitially}
\alias{gtkWindowGetKeysChangedHandler}
\alias{gtkWindowGetMaximizeInitially}
\alias{gtkWindowGetNeedDefaultPosition}
\alias{gtkWindowGetNeedDefaultSize}
\alias{gtkWindowGetStickInitially}
\alias{gtkWindowGetTransientParent}
\alias{gtkWindowGetWmRole}
\alias{gtkWindowGetWmclassClass}
\alias{gtkWindowGetWmclassName}
\alias{pangoAnalysisGetExtraAttrs}
\alias{pangoAnalysisGetFont}
\alias{pangoAnalysisGetLanguage}
\alias{pangoAnalysisGetLevel}
\alias{pangoAttrColorGetColor}
\alias{pangoAttrFloatGetValue}
\alias{pangoAttrFontDescGetDesc}
\alias{pangoAttrIntGetValue}
\alias{pangoAttrLanguageGetValue}
\alias{pangoAttrShapeGetInkRect}
\alias{pangoAttrShapeGetLogicalRect}
\alias{pangoAttrSizeGetAbsolute}
\alias{pangoAttrSizeGetSize}
\alias{pangoAttrStringGetValue}
\alias{pangoAttributeGetEndIndex}
\alias{pangoAttributeGetKlass}
\alias{pangoAttributeGetStartIndex}
\alias{pangoColorGetBlue}
\alias{pangoColorGetGreen}
\alias{pangoColorGetRed}
\alias{pangoGlyphGeometryGetWidth}
\alias{pangoGlyphGeometryGetXOffset}
\alias{pangoGlyphGeometryGetYOffset}
\alias{pangoGlyphInfoGetAttr}
\alias{pangoGlyphInfoGetGeometry}
\alias{pangoGlyphInfoGetGlyph}
\alias{pangoGlyphItemGetGlyphs}
\alias{pangoGlyphItemGetItem}
\alias{pangoGlyphStringGetGlyphs}
\alias{pangoGlyphStringGetLogClusters}
\alias{pangoGlyphStringGetNumGlyphs}
\alias{pangoGlyphVisAttrGetIsClusterStart}
\alias{pangoItemGetAnalysis}
\alias{pangoItemGetLength}
\alias{pangoItemGetNumChars}
\alias{pangoItemGetOffset}
\alias{pangoLayoutLineGetIsParagraphStart}
\alias{pangoLayoutLineGetLayout}
\alias{pangoLayoutLineGetLength}
\alias{pangoLayoutLineGetResolvedDir}
\alias{pangoLayoutLineGetRuns}
\alias{pangoLayoutLineGetStartIndex}
\alias{pangoLogAttrGetBackspaceDeletesCharacter}
\alias{pangoLogAttrGetIsCharBreak}
\alias{pangoLogAttrGetIsCursorPosition}
\alias{pangoLogAttrGetIsLineBreak}
\alias{pangoLogAttrGetIsMandatoryBreak}
\alias{pangoLogAttrGetIsSentenceBoundary}
\alias{pangoLogAttrGetIsSentenceEnd}
\alias{pangoLogAttrGetIsSentenceStart}
\alias{pangoLogAttrGetIsWhite}
\alias{pangoLogAttrGetIsWordEnd}
\alias{pangoLogAttrGetIsWordStart}
\alias{gdkImageGetBitsPerPixel}
\alias{gdkImageGetBpl}
\alias{gdkImageGetBpp}
\alias{gdkImageGetByteOrder}
\alias{gdkImageGetDepth}
\alias{gdkImageGetHeight}
\alias{gdkImageGetMem}
\alias{gdkImageGetVisual}
\alias{gdkImageGetWidth}
\alias{gFileAttributeInfoListGetInfos}
\alias{gFileAttributeInfoListGetNInfos}
\name{getters}
\title{Getters}
\description{This function gets the \code{Field} from the \code{Structure} in its name,
following the format \code{StructureGetField}. You should use the \code{\link{[[.RGtkObject}} function instead 
of this one.}
\author{Michael Lawrence}
\keyword{interface}
\keyword{internal}
