\alias{gFileAttributeInfoListNew}
\name{gFileAttributeInfoListNew}
\title{gFileAttributeInfoListNew}
\description{Creates a new file attribute info list.}
\usage{gFileAttributeInfoListNew()}
\value{[\code{\link{GFileAttributeInfoList}}] a \code{\link{GFileAttributeInfoList}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
