\alias{gFileAttributeMatcherEnumerateNext}
\name{gFileAttributeMatcherEnumerateNext}
\title{gFileAttributeMatcherEnumerateNext}
\description{Gets the next matched attribute from a \code{\link{GFileAttributeMatcher}}.}
\usage{gFileAttributeMatcherEnumerateNext(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFileAttributeMatcher}}.}}
\value{[char] a string containing the next attribute or \code{NULL} if
no more attribute exist.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
