\alias{gFileEnumeratorGetContainer}
\name{gFileEnumeratorGetContainer}
\title{gFileEnumeratorGetContainer}
\description{Get the \code{\link{GFile}} container which is being enumerated.}
\usage{gFileEnumeratorGetContainer(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFileEnumerator}}}}
\details{Since 2.18}
\value{[\code{\link{GFile}}] the \code{\link{GFile}} which is being enumerated.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
