\alias{gFileEnumeratorHasPending}
\name{gFileEnumeratorHasPending}
\title{gFileEnumeratorHasPending}
\description{Checks if the file enumerator has pending operations.}
\usage{gFileEnumeratorHasPending(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFileEnumerator}}.}}
\value{[logical] \code{TRUE} if the \code{enumerator} has pending operations.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
