\alias{gFileEnumeratorIsClosed}
\name{gFileEnumeratorIsClosed}
\title{gFileEnumeratorIsClosed}
\description{Checks if the file enumerator has been closed.}
\usage{gFileEnumeratorIsClosed(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFileEnumerator}}.}}
\value{[logical] \code{TRUE} if the \code{enumerator} is closed.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
