\alias{gFileEnumeratorSetPending}
\name{gFileEnumeratorSetPending}
\title{gFileEnumeratorSetPending}
\description{Sets the file enumerator as having pending operations.}
\usage{gFileEnumeratorSetPending(object, pending)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileEnumerator}}.}
\item{\verb{pending}}{a boolean value.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
