\alias{gFileGetUri}
\name{gFileGetUri}
\title{gFileGetUri}
\description{Gets the URI for the \code{file}.}
\usage{gFileGetUri(object)}
\arguments{\item{\verb{object}}{input \code{\link{GFile}}.}}
\details{This call does no blocking i/o.}
\value{[char] a string containing the \code{\link{GFile}}'s URI.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
