\alias{gFileIconGetFile}
\name{gFileIconGetFile}
\title{gFileIconGetFile}
\description{Gets the \code{\link{GFile}} associated with the given \code{icon}.}
\usage{gFileIconGetFile(object)}
\arguments{\item{\verb{object}}{a \code{\link{GIcon}}.}}
\value{[\code{\link{GFile}}] a \code{\link{GFile}}, or \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
