\alias{gFileIconNew}
\name{gFileIconNew}
\title{gFileIconNew}
\description{Creates a new icon for a file.}
\usage{gFileIconNew(file)}
\arguments{\item{\verb{file}}{a \code{\link{GFile}}.}}
\value{[\code{\link{GIcon}}] a \code{\link{GIcon}} for the given \code{file}, or \code{NULL} on error.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
