\alias{gFileInfoClearStatus}
\name{gFileInfoClearStatus}
\title{gFileInfoClearStatus}
\description{Clears the status information from \code{info}.}
\usage{gFileInfoClearStatus(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFileInfo}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
