\alias{gFileInfoCopyInto}
\name{gFileInfoCopyInto}
\title{gFileInfoCopyInto}
\description{Copies all of the \verb{GFileAttribute}s from \code{src.info} to \code{dest.info}.}
\usage{gFileInfoCopyInto(object, dest.info)}
\arguments{
\item{\verb{object}}{source to copy attributes from.}
\item{\verb{dest.info}}{destination to copy attributes to.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
