\alias{gFileInfoDup}
\name{gFileInfoDup}
\title{gFileInfoDup}
\description{Duplicates a file info structure.}
\usage{gFileInfoDup(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFileInfo}}.}}
\value{[\code{\link{GFileInfo}}] a duplicate \code{\link{GFileInfo}} of \code{other}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
