\alias{gFileInfoGetAttributeBoolean}
\name{gFileInfoGetAttributeBoolean}
\title{gFileInfoGetAttributeBoolean}
\description{Gets the value of a boolean attribute. If the attribute does not
contain a boolean value, \code{FALSE} will be returned.}
\usage{gFileInfoGetAttributeBoolean(object, attribute)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{attribute}}{a file attribute key.}
}
\value{[logical] the boolean value contained within the attribute.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
