\alias{gFileInfoGetContentType}
\name{gFileInfoGetContentType}
\title{gFileInfoGetContentType}
\description{Gets the file's content type.}
\usage{gFileInfoGetContentType(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFileInfo}}.}}
\value{[char] a string containing the file's content type.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
