\alias{gFileInfoGetDisplayName}
\name{gFileInfoGetDisplayName}
\title{gFileInfoGetDisplayName}
\description{Gets a display name for a file.}
\usage{gFileInfoGetDisplayName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFileInfo}}.}}
\value{[char] a string containing the display name.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
