\alias{gFileInfoGetEditName}
\name{gFileInfoGetEditName}
\title{gFileInfoGetEditName}
\description{Gets the edit name for a file.}
\usage{gFileInfoGetEditName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFileInfo}}.}}
\value{[char] a string containing the edit name.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
