\alias{gFileInfoGetEtag}
\name{gFileInfoGetEtag}
\title{gFileInfoGetEtag}
\description{Gets the entity tag for a given 
\code{\link{GFileInfo}}. See \code{G_FILE_ATTRIBUTE_ETAG_VALUE}.}
\usage{gFileInfoGetEtag(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFileInfo}}.}}
\value{[char] a string containing the value of the "etag:value" attribute.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
