\alias{gFileInfoGetIcon}
\name{gFileInfoGetIcon}
\title{gFileInfoGetIcon}
\description{Gets the icon for a file.}
\usage{gFileInfoGetIcon(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFileInfo}}.}}
\value{[\code{\link{GIcon}}] \code{\link{GIcon}} for the given \code{info}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
