\alias{gFileInfoGetIsBackup}
\name{gFileInfoGetIsBackup}
\title{gFileInfoGetIsBackup}
\description{Checks if a file is a backup file.}
\usage{gFileInfoGetIsBackup(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFileInfo}}.}}
\value{[logical] \code{TRUE} if file is a backup file, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
