\alias{gFileInfoGetIsHidden}
\name{gFileInfoGetIsHidden}
\title{gFileInfoGetIsHidden}
\description{Checks if a file is hidden.}
\usage{gFileInfoGetIsHidden(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFileInfo}}.}}
\value{[logical] \code{TRUE} if the file is a hidden file, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
