\alias{gFileInfoGetIsSymlink}
\name{gFileInfoGetIsSymlink}
\title{gFileInfoGetIsSymlink}
\description{Checks if a file is a symlink.}
\usage{gFileInfoGetIsSymlink(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFileInfo}}.}}
\value{[logical] \code{TRUE} if the given \code{info} is a symlink.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
