\alias{gFileInfoGetModificationTime}
\name{gFileInfoGetModificationTime}
\title{gFileInfoGetModificationTime}
\description{Gets the modification time of the current \code{info} and sets it
in \code{result}.}
\usage{gFileInfoGetModificationTime(object, result)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{result}}{a \code{\link{GTimeVal}}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
