\alias{gFileInfoGetName}
\name{gFileInfoGetName}
\title{gFileInfoGetName}
\description{Gets the name for a file.}
\usage{gFileInfoGetName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFileInfo}}.}}
\value{[char] a string containing the file name.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
