\alias{gFileInfoGetSize}
\name{gFileInfoGetSize}
\title{gFileInfoGetSize}
\description{Gets the file's size.}
\usage{gFileInfoGetSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFileInfo}}.}}
\value{[numeric] a \verb{numeric} containing the file's size.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
