\alias{gFileInfoGetSortOrder}
\name{gFileInfoGetSortOrder}
\title{gFileInfoGetSortOrder}
\description{Gets the value of the sort_order attribute from the \code{\link{GFileInfo}}.
See \code{G_FILE_ATTRIBUTE_STANDARD_SORT_ORDER}.}
\usage{gFileInfoGetSortOrder(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFileInfo}}.}}
\value{[integer] a \verb{integer} containing the value of the "standard::sort_order" attribute.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
