\alias{gFileInfoGetSymlinkTarget}
\name{gFileInfoGetSymlinkTarget}
\title{gFileInfoGetSymlinkTarget}
\description{Gets the symlink target for a given \code{\link{GFileInfo}}.}
\usage{gFileInfoGetSymlinkTarget(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFileInfo}}.}}
\value{[char] a string containing the symlink target.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
