\alias{gFileInfoHasAttribute}
\name{gFileInfoHasAttribute}
\title{gFileInfoHasAttribute}
\description{Checks if a file info structure has an attribute named \code{attribute}.}
\usage{gFileInfoHasAttribute(object, attribute)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{attribute}}{a file attribute key.}
}
\value{[logical] \code{TRUE} if \code{Ginfo} has an attribute named \code{attribute},
\code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
