\alias{gFileInfoHasNamespace}
\name{gFileInfoHasNamespace}
\title{gFileInfoHasNamespace}
\description{Checks if a file info structure has an attribute in the
specified \code{name.space}.}
\usage{gFileInfoHasNamespace(object, name.space)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{name.space}}{a file attribute namespace.}
}
\details{Since 2.22}
\value{[logical] \code{TRUE} if \code{Ginfo} has an attribute in \code{name.space},
\code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
