\alias{gFileInfoRemoveAttribute}
\name{gFileInfoRemoveAttribute}
\title{gFileInfoRemoveAttribute}
\description{Removes all cases of \code{attribute} from \code{info} if it exists.}
\usage{gFileInfoRemoveAttribute(object, attribute)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{attribute}}{a file attribute key.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
