\alias{gFileInfoSetAttribute}
\name{gFileInfoSetAttribute}
\title{gFileInfoSetAttribute}
\description{Sets the \code{attribute} to contain the given value, if possible.}
\usage{gFileInfoSetAttribute(object, attribute, type, value.p)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{attribute}}{a file attribute key.}
\item{\verb{type}}{a \code{\link{GFileAttributeType}}}
\item{\verb{value.p}}{pointer to the value}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
