\alias{gFileInfoSetAttributeBoolean}
\name{gFileInfoSetAttributeBoolean}
\title{gFileInfoSetAttributeBoolean}
\description{Sets the \code{attribute} to contain the given \code{attr.value}, 
if possible.}
\usage{gFileInfoSetAttributeBoolean(object, attribute, attr.value)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{attribute}}{a file attribute key.}
\item{\verb{attr.value}}{a boolean value.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
