\alias{gFileInfoSetAttributeByteString}
\name{gFileInfoSetAttributeByteString}
\title{gFileInfoSetAttributeByteString}
\description{Sets the \code{attribute} to contain the given \code{attr.value}, 
if possible.}
\usage{gFileInfoSetAttributeByteString(object, attribute, attr.value)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{attribute}}{a file attribute key.}
\item{\verb{attr.value}}{a byte string.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
