\alias{gFileInfoSetAttributeInt32}
\name{gFileInfoSetAttributeInt32}
\title{gFileInfoSetAttributeInt32}
\description{Sets the \code{attribute} to contain the given \code{attr.value}, 
if possible.}
\usage{gFileInfoSetAttributeInt32(object, attribute, attr.value)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{attribute}}{a file attribute key.}
\item{\verb{attr.value}}{a signed 32-bit integer}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
