\alias{gFileInfoSetAttributeInt64}
\name{gFileInfoSetAttributeInt64}
\title{gFileInfoSetAttributeInt64}
\description{Sets the \code{attribute} to contain the given \code{attr.value}, 
if possible.}
\usage{gFileInfoSetAttributeInt64(object, attribute, attr.value)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{attribute}}{attribute name to set.}
\item{\verb{attr.value}}{int64 value to set attribute to.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
