\alias{gFileInfoSetAttributeMask}
\name{gFileInfoSetAttributeMask}
\title{gFileInfoSetAttributeMask}
\description{Sets \code{mask} on \code{info} to match specific attribute types.}
\usage{gFileInfoSetAttributeMask(object, mask)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{mask}}{a \code{\link{GFileAttributeMatcher}}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
