\alias{gFileInfoSetAttributeString}
\name{gFileInfoSetAttributeString}
\title{gFileInfoSetAttributeString}
\description{Sets the \code{attribute} to contain the given \code{attr.value}, 
if possible.}
\usage{gFileInfoSetAttributeString(object, attribute, attr.value)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{attribute}}{a file attribute key.}
\item{\verb{attr.value}}{a string.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
