\alias{gFileInfoSetAttributeStringv}
\name{gFileInfoSetAttributeStringv}
\title{gFileInfoSetAttributeStringv}
\description{Sets the \code{attribute} to contain the given \code{attr.value},
if possible.}
\usage{gFileInfoSetAttributeStringv(object, attribute, attr.value)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{attribute}}{a file attribute key.}
\item{\verb{attr.value}}{a \code{NULL} terminated string list}
}
\details{Sinze: 2.22}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
