\alias{gFileInfoSetAttributeUint64}
\name{gFileInfoSetAttributeUint64}
\title{gFileInfoSetAttributeUint64}
\description{Sets the \code{attribute} to contain the given \code{attr.value}, 
if possible.}
\usage{gFileInfoSetAttributeUint64(object, attribute, attr.value)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{attribute}}{a file attribute key.}
\item{\verb{attr.value}}{an unsigned 64-bit integer.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
