\alias{gFileInfoSetContentType}
\name{gFileInfoSetContentType}
\title{gFileInfoSetContentType}
\description{Sets the content type attribute for a given \code{\link{GFileInfo}}.
See \code{G_FILE_ATTRIBUTE_STANDARD_CONTENT_TYPE}.}
\usage{gFileInfoSetContentType(object, content.type)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{content.type}}{a content type. See \verb{GContentType}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
