\alias{gFileInfoSetDisplayName}
\name{gFileInfoSetDisplayName}
\title{gFileInfoSetDisplayName}
\description{Sets the display name for the current \code{\link{GFileInfo}}.
See \code{G_FILE_ATTRIBUTE_STANDARD_DISPLAY_NAME}.}
\usage{gFileInfoSetDisplayName(object, display.name)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{display.name}}{a string containing a display name.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
