\alias{gFileInfoSetEditName}
\name{gFileInfoSetEditName}
\title{gFileInfoSetEditName}
\description{Sets the edit name for the current file.
See \code{G_FILE_ATTRIBUTE_STANDARD_EDIT_NAME}.}
\usage{gFileInfoSetEditName(object, edit.name)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{edit.name}}{a string containing an edit name.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
