\alias{gFileInfoSetFileType}
\name{gFileInfoSetFileType}
\title{gFileInfoSetFileType}
\description{Sets the file type in a \code{\link{GFileInfo}} to \code{type}.
See \code{G_FILE_ATTRIBUTE_STANDARD_TYPE}.}
\usage{gFileInfoSetFileType(object, type)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{type}}{a \code{\link{GFileType}}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
