\alias{gFileInfoSetIcon}
\name{gFileInfoSetIcon}
\title{gFileInfoSetIcon}
\description{Sets the icon for a given \code{\link{GFileInfo}}. 
See \code{G_FILE_ATTRIBUTE_STANDARD_ICON}.}
\usage{gFileInfoSetIcon(object, icon)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{icon}}{a \code{\link{GIcon}}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
