\alias{gFileInfoSetIsHidden}
\name{gFileInfoSetIsHidden}
\title{gFileInfoSetIsHidden}
\description{Sets the "is_hidden" attribute in a \code{\link{GFileInfo}} according to \code{is.symlink}.
See \code{G_FILE_ATTRIBUTE_STANDARD_IS_HIDDEN}.}
\usage{gFileInfoSetIsHidden(object, is.hidden)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{is.hidden}}{a \verb{logical}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
