\alias{gFileInfoSetIsSymlink}
\name{gFileInfoSetIsSymlink}
\title{gFileInfoSetIsSymlink}
\description{Sets the "is_symlink" attribute in a \code{\link{GFileInfo}} according to \code{is.symlink}.
See \code{G_FILE_ATTRIBUTE_STANDARD_IS_SYMLINK}.}
\usage{gFileInfoSetIsSymlink(object, is.symlink)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{is.symlink}}{a \verb{logical}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
