\alias{gFileInfoSetModificationTime}
\name{gFileInfoSetModificationTime}
\title{gFileInfoSetModificationTime}
\description{Sets the \code{G_FILE_ATTRIBUTE_TIME_MODIFIED} attribute in the file
info to the given time value.}
\usage{gFileInfoSetModificationTime(object, mtime)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{mtime}}{a \code{\link{GTimeVal}}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
