\alias{gFileInfoSetName}
\name{gFileInfoSetName}
\title{gFileInfoSetName}
\description{Sets the name attribute for the current \code{\link{GFileInfo}}. 
See \code{G_FILE_ATTRIBUTE_STANDARD_NAME}.}
\usage{gFileInfoSetName(object, name)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{name}}{a string containing a name.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
