\alias{gFileInfoSetSize}
\name{gFileInfoSetSize}
\title{gFileInfoSetSize}
\description{Sets the \code{G_FILE_ATTRIBUTE_STANDARD_SIZE} attribute in the file info 
to the given size.}
\usage{gFileInfoSetSize(object, size)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{size}}{a \verb{numeric} containing the file's size.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
