\alias{gFileInfoSetSortOrder}
\name{gFileInfoSetSortOrder}
\title{gFileInfoSetSortOrder}
\description{Sets the sort order attribute in the file info structure. See 
\code{G_FILE_ATTRIBUTE_STANDARD_SORT_ORDER}.}
\usage{gFileInfoSetSortOrder(object, sort.order)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{sort.order}}{a sort order integer.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
