\alias{gFileInfoSetSymlinkTarget}
\name{gFileInfoSetSymlinkTarget}
\title{gFileInfoSetSymlinkTarget}
\description{Sets the \code{G_FILE_ATTRIBUTE_STANDARD_SYMLINK_TARGET} attribute in the file info 
to the given symlink target.}
\usage{gFileInfoSetSymlinkTarget(object, symlink.target)}
\arguments{
\item{\verb{object}}{a \code{\link{GFileInfo}}.}
\item{\verb{symlink.target}}{a static string containing a path to a symlink target.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
