\alias{gFileInfoUnsetAttributeMask}
\name{gFileInfoUnsetAttributeMask}
\title{gFileInfoUnsetAttributeMask}
\description{Unsets a mask set by \code{\link{gFileInfoSetAttributeMask}}, if one
is set.}
\usage{gFileInfoUnsetAttributeMask(object)}
\arguments{\item{\verb{object}}{\code{\link{GFileInfo}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
