\alias{gFileMonitorCancel}
\name{gFileMonitorCancel}
\title{gFileMonitorCancel}
\description{Cancels a file monitor.}
\usage{gFileMonitorCancel(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFileMonitor}}.}}
\value{[logical] \code{TRUE} if monitor was cancelled.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
