\alias{gFileMonitorIsCancelled}
\name{gFileMonitorIsCancelled}
\title{gFileMonitorIsCancelled}
\description{Returns whether the monitor is canceled.}
\usage{gFileMonitorIsCancelled(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFileMonitor}}}}
\value{[logical] \code{TRUE} if monitor is canceled. \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
