\alias{gFilenameCompleterGetCompletions}
\name{gFilenameCompleterGetCompletions}
\title{gFilenameCompleterGetCompletions}
\description{Gets a list of completion strings for a given initial text.}
\usage{gFilenameCompleterGetCompletions(object, initial.text)}
\arguments{
\item{\verb{object}}{the filename completer.}
\item{\verb{initial.text}}{text to be completed.}
}
\value{[char] array of strings with possible completions for \code{initial.text}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
