\alias{gFilenameCompleterGetCompletionSuffix}
\name{gFilenameCompleterGetCompletionSuffix}
\title{gFilenameCompleterGetCompletionSuffix}
\description{Obtains a completion for \code{initial.text} from \code{completer}.}
\usage{gFilenameCompleterGetCompletionSuffix(object, initial.text)}
\arguments{
\item{\verb{object}}{the filename completer.}
\item{\verb{initial.text}}{text to be completed.}
}
\value{[char] a completed string, or \code{NULL} if no completion exists.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
