\alias{gFilenameCompleterNew}
\name{gFilenameCompleterNew}
\title{gFilenameCompleterNew}
\description{Creates a new filename completer.}
\usage{gFilenameCompleterNew()}
\value{[\code{\link{GFilenameCompleter}}] a \code{\link{GFilenameCompleter}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
