\alias{gFilenameCompleterSetDirsOnly}
\name{gFilenameCompleterSetDirsOnly}
\title{gFilenameCompleterSetDirsOnly}
\description{If \code{dirs.only} is \code{TRUE}, \code{completer} will only 
complete directory names, and not file names.}
\usage{gFilenameCompleterSetDirsOnly(object, dirs.only)}
\arguments{
\item{\verb{object}}{the filename completer.}
\item{\verb{dirs.only}}{a \verb{logical}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
