\alias{gFilterInputStreamGetBaseStream}
\name{gFilterInputStreamGetBaseStream}
\title{gFilterInputStreamGetBaseStream}
\description{Gets the base stream for the filter stream.}
\usage{gFilterInputStreamGetBaseStream(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFilterInputStream}}.}}
\value{[\code{\link{GInputStream}}] a \code{\link{GInputStream}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
