\alias{gFilterInputStreamGetCloseBaseStream}
\name{gFilterInputStreamGetCloseBaseStream}
\title{gFilterInputStreamGetCloseBaseStream}
\description{Returns whether the base stream will be closed when \code{stream} is
closed.}
\usage{gFilterInputStreamGetCloseBaseStream(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFilterInputStream}}.}}
\value{[logical] \code{TRUE} if the base stream will be closed.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
