\alias{gFilterInputStreamSetCloseBaseStream}
\name{gFilterInputStreamSetCloseBaseStream}
\title{gFilterInputStreamSetCloseBaseStream}
\description{Sets whether the base stream will be closed when \code{stream} is closed.}
\usage{gFilterInputStreamSetCloseBaseStream(object, close.base)}
\arguments{
\item{\verb{object}}{a \code{\link{GFilterInputStream}}.}
\item{\verb{close.base}}{\code{TRUE} to close the base stream.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
