\alias{gFilterOutputStreamGetBaseStream}
\name{gFilterOutputStreamGetBaseStream}
\title{gFilterOutputStreamGetBaseStream}
\description{Gets the base stream for the filter stream.}
\usage{gFilterOutputStreamGetBaseStream(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFilterOutputStream}}.}}
\value{[\code{\link{GOutputStream}}] a \code{\link{GOutputStream}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
