\alias{gFilterOutputStreamGetCloseBaseStream}
\name{gFilterOutputStreamGetCloseBaseStream}
\title{gFilterOutputStreamGetCloseBaseStream}
\description{Returns whether the base stream will be closed when \code{stream} is
closed.}
\usage{gFilterOutputStreamGetCloseBaseStream(object)}
\arguments{\item{\verb{object}}{a \code{\link{GFilterOutputStream}}.}}
\value{[logical] \code{TRUE} if the base stream will be closed.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
