\alias{gFilterOutputStreamSetCloseBaseStream}
\name{gFilterOutputStreamSetCloseBaseStream}
\title{gFilterOutputStreamSetCloseBaseStream}
\description{Sets whether the base stream will be closed when \code{stream} is closed.}
\usage{gFilterOutputStreamSetCloseBaseStream(object, close.base)}
\arguments{
\item{\verb{object}}{a \code{\link{GFilterOutputStream}}.}
\item{\verb{close.base}}{\code{TRUE} to close the base stream.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
