\alias{gIconEqual}
\name{gIconEqual}
\title{gIconEqual}
\description{Checks if two icons are equal.}
\usage{gIconEqual(object, icon2)}
\arguments{
\item{\verb{object}}{pointer to the first \code{\link{GIcon}}.}
\item{\verb{icon2}}{pointer to the second \code{\link{GIcon}}.}
}
\value{[logical] \code{TRUE} if \code{icon1} is equal to \code{icon2}. \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
