\alias{gIconHash}
\name{gIconHash}
\title{gIconHash}
\description{Gets a hash for an icon.}
\usage{gIconHash(icon)}
\arguments{\item{\verb{icon}}{\verb{R object} to an icon object.}}
\value{[numeric] a \verb{numeric} containing a hash for the \code{icon}, suitable for
use in a \verb{GHashTable} or similar data structure.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
