\alias{gInetAddressGetFamily}
\name{gInetAddressGetFamily}
\title{gInetAddressGetFamily}
\description{Gets \code{address}'s family}
\usage{gInetAddressGetFamily(object)}
\arguments{\item{\verb{object}}{a \code{\link{GInetAddress}}}}
\details{Since 2.22}
\value{[\code{\link{GSocketFamily}}] \code{address}'s family}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
