\alias{gInetAddressGetIsAny}
\name{gInetAddressGetIsAny}
\title{gInetAddressGetIsAny}
\description{Tests whether \code{address} is the "any" address for its family.}
\usage{gInetAddressGetIsAny(object)}
\arguments{\item{\verb{object}}{a \code{\link{GInetAddress}}}}
\details{Since 2.22}
\value{[logical] \code{TRUE} if \code{address} is the "any" address for its family.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
