\alias{gInetAddressGetIsLinkLocal}
\name{gInetAddressGetIsLinkLocal}
\title{gInetAddressGetIsLinkLocal}
\description{Tests whether \code{address} is a link-local address (that is, if it
identifies a host on a local network that is not connected to the
Internet).}
\usage{gInetAddressGetIsLinkLocal(object)}
\arguments{\item{\verb{object}}{a \code{\link{GInetAddress}}}}
\details{Since 2.22}
\value{[logical] \code{TRUE} if \code{address} is a link-local address.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
