\alias{gInetAddressGetIsLoopback}
\name{gInetAddressGetIsLoopback}
\title{gInetAddressGetIsLoopback}
\description{Tests whether \code{address} is the loopback address for its family.}
\usage{gInetAddressGetIsLoopback(object)}
\arguments{\item{\verb{object}}{a \code{\link{GInetAddress}}}}
\details{Since 2.22}
\value{[logical] \code{TRUE} if \code{address} is the loopback address for its family.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
