\alias{gInetAddressGetIsMcGlobal}
\name{gInetAddressGetIsMcGlobal}
\title{gInetAddressGetIsMcGlobal}
\description{Tests whether \code{address} is a global multicast address.}
\usage{gInetAddressGetIsMcGlobal(object)}
\arguments{\item{\verb{object}}{a \code{\link{GInetAddress}}}}
\details{Since 2.22}
\value{[logical] \code{TRUE} if \code{address} is a global multicast address.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
