\alias{gInetAddressGetIsMcLinkLocal}
\name{gInetAddressGetIsMcLinkLocal}
\title{gInetAddressGetIsMcLinkLocal}
\description{Tests whether \code{address} is a link-local multicast address.}
\usage{gInetAddressGetIsMcLinkLocal(object)}
\arguments{\item{\verb{object}}{a \code{\link{GInetAddress}}}}
\details{Since 2.22}
\value{[logical] \code{TRUE} if \code{address} is a link-local multicast address.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
