\alias{gInetAddressGetIsMcNodeLocal}
\name{gInetAddressGetIsMcNodeLocal}
\title{gInetAddressGetIsMcNodeLocal}
\description{Tests whether \code{address} is a node-local multicast address.}
\usage{gInetAddressGetIsMcNodeLocal(object)}
\arguments{\item{\verb{object}}{a \code{\link{GInetAddress}}}}
\details{Since 2.22}
\value{[logical] \code{TRUE} if \code{address} is a node-local multicast address.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
