\alias{gInetAddressGetIsMcOrgLocal}
\name{gInetAddressGetIsMcOrgLocal}
\title{gInetAddressGetIsMcOrgLocal}
\description{Tests whether \code{address} is an organization-local multicast address.}
\usage{gInetAddressGetIsMcOrgLocal(object)}
\arguments{\item{\verb{object}}{a \code{\link{GInetAddress}}}}
\details{Since 2.22}
\value{[logical] \code{TRUE} if \code{address} is an organization-local multicast address.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
