\alias{gInetAddressGetIsMcSiteLocal}
\name{gInetAddressGetIsMcSiteLocal}
\title{gInetAddressGetIsMcSiteLocal}
\description{Tests whether \code{address} is a site-local multicast address.}
\usage{gInetAddressGetIsMcSiteLocal(object)}
\arguments{\item{\verb{object}}{a \code{\link{GInetAddress}}}}
\details{Since 2.22}
\value{[logical] \code{TRUE} if \code{address} is a site-local multicast address.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
