\alias{gInetAddressGetIsMulticast}
\name{gInetAddressGetIsMulticast}
\title{gInetAddressGetIsMulticast}
\description{Tests whether \code{address} is a multicast address.}
\usage{gInetAddressGetIsMulticast(object)}
\arguments{\item{\verb{object}}{a \code{\link{GInetAddress}}}}
\details{Since 2.22}
\value{[logical] \code{TRUE} if \code{address} is a multicast address.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
