\alias{gInetAddressGetNativeSize}
\name{gInetAddressGetNativeSize}
\title{gInetAddressGetNativeSize}
\description{Gets the size of the native raw binary address for \code{address}. This
is the size of the data that you get from \code{\link{gInetAddressToBytes}}.}
\usage{gInetAddressGetNativeSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GInetAddress}}}}
\details{Since 2.22}
\value{[numeric] the number of bytes used for the native version of \code{address}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
