\alias{gInetAddressNewAny}
\name{gInetAddressNewAny}
\title{gInetAddressNewAny}
\description{Creates a \code{\link{GInetAddress}} for the "any" address (unassigned/"don't
care") for \code{family}.}
\usage{gInetAddressNewAny(family)}
\arguments{\item{\verb{family}}{the address family}}
\details{Since 2.22}
\value{[\code{\link{GInetAddress}}] a new \code{\link{GInetAddress}} corresponding to the "any" address
for \code{family}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
