\alias{gInetAddressNewLoopback}
\name{gInetAddressNewLoopback}
\title{gInetAddressNewLoopback}
\description{Creates a \code{\link{GInetAddress}} for the loopback address for \code{family}.}
\usage{gInetAddressNewLoopback(family)}
\arguments{\item{\verb{family}}{the address family}}
\details{Since 2.22}
\value{[\code{\link{GInetAddress}}] a new \code{\link{GInetAddress}} corresponding to the loopback address
for \code{family}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
