\alias{gInetAddressToString}
\name{gInetAddressToString}
\title{gInetAddressToString}
\description{Converts \code{address} to string form.}
\usage{gInetAddressToString(object)}
\arguments{\item{\verb{object}}{a \code{\link{GInetAddress}}}}
\details{Since 2.22}
\value{[character] a representation of \code{address} as a string,}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
