\alias{gInetSocketAddressGetAddress}
\name{gInetSocketAddressGetAddress}
\title{gInetSocketAddressGetAddress}
\description{Gets \code{address}'s \code{\link{GInetAddress}}.}
\usage{gInetSocketAddressGetAddress(object)}
\arguments{\item{\verb{object}}{a \code{\link{GInetSocketAddress}}}}
\details{Since 2.22}
\value{[\code{\link{GInetAddress}}] the \code{\link{GInetAddress}} for \code{address},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
