\alias{gInetSocketAddressGetPort}
\name{gInetSocketAddressGetPort}
\title{gInetSocketAddressGetPort}
\description{Gets \code{address}'s port.}
\usage{gInetSocketAddressGetPort(object)}
\arguments{\item{\verb{object}}{a \code{\link{GInetSocketAddress}}}}
\details{Since 2.22}
\value{[integer] the port for \code{address}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
