\alias{gInetSocketAddressNew}
\name{gInetSocketAddressNew}
\title{gInetSocketAddressNew}
\description{Creates a new \code{\link{GInetSocketAddress}} for \code{address} and \code{port}.}
\usage{gInetSocketAddressNew(address, port)}
\arguments{
\item{\verb{address}}{a \code{\link{GInetAddress}}}
\item{\verb{port}}{a port number}
}
\details{Since 2.22}
\value{[\code{\link{GSocketAddress}}] a new \code{\link{GInetSocketAddress}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
