\alias{gInputStreamClearPending}
\name{gInputStreamClearPending}
\title{gInputStreamClearPending}
\description{Clears the pending flag on \code{stream}.}
\usage{gInputStreamClearPending(object)}
\arguments{\item{\verb{object}}{input stream}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
