\alias{gInputStreamHasPending}
\name{gInputStreamHasPending}
\title{gInputStreamHasPending}
\description{Checks if an input stream has pending actions.}
\usage{gInputStreamHasPending(object)}
\arguments{\item{\verb{object}}{input stream.}}
\value{[logical] \code{TRUE} if \code{stream} has pending actions.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
