\alias{gInputStreamIsClosed}
\name{gInputStreamIsClosed}
\title{gInputStreamIsClosed}
\description{Checks if an input stream is closed.}
\usage{gInputStreamIsClosed(object)}
\arguments{\item{\verb{object}}{input stream.}}
\value{[logical] \code{TRUE} if the stream is closed.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
