\alias{gio-Desktop-file-based-GAppInfo}
\name{gio-Desktop-file-based-GAppInfo}
\title{Desktop file based GAppInfo}
\description{Application information from desktop files}
\section{Hierarchy}{\preformatted{
GObject
   +----GDesktopAppInfo
GInterface
   +----GDesktopAppInfoLookup
}}
\section{Interfaces}{GDesktopAppInfo implements
 \verb{\link{GAppInfo}}.}
\section{Detailed Description}{\verb{GDesktopAppInfo} is an implementation of \verb{\link{GAppInfo}} based on
desktop files.
  
Note that \file{<gio/gdesktopappinfo.h>} belongs to 
the UNIX-specific GIO interfaces, thus you have to use the 
\file{gio-unix-2.0.pc} pkg-config file when using it.}
\references{\url{https://developer-old.gnome.org/gio/stable/gio-Desktop-file-based-GAppInfo.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
