\alias{GIO}
\name{GIO}
\title{GIO}
\description{GIO is a modern, easy-to-use VFS API}
\details{
The RGtk binding to the GIO library consists of the following components:
\describe{
\item{\link{gio-Extension-Points}}{Extension Points}
\item{\link{GAppInfo}}{Application information and launch contexts}
\item{\link{GAsyncInitable}}{Asynchronously failable object initialization interface}
\item{\link{GAsyncResult}}{Asynchronous Function Results}
\item{\link{GBufferedInputStream}}{Buffered Input Stream}
\item{\link{GBufferedOutputStream}}{Buffered Output Stream}
\item{\link{GCancellable}}{Thread-safe Operation Cancellation Stack}
\item{\link{gio-GContentType}}{Platform-specific content typing}
\item{\link{GDataInputStream}}{Data Input Stream}
\item{\link{GDataOutputStream}}{Data Output Stream}
\item{\link{GDrive}}{Drive management}
\item{\link{GEmblem}}{An object for emblems}
\item{\link{GEmblemedIcon}}{Icon with emblems}
\item{\link{GFile}}{File and Directory Handling}
\item{\link{gio-GFileAttribute}}{Key-Value Paired File Attributes}
\item{\link{GFileEnumerator}}{Enumerated Files Routines}
\item{\link{GFileIcon}}{Icons pointing to an image file}
\item{\link{GFileInfo}}{File Information and Attributes}
\item{\link{GFileInputStream}}{File input streaming operations}
\item{\link{GFileIOStream}}{File read and write streaming operations}
\item{\link{GFileMonitor}}{File Monitor}
\item{\link{GFilenameCompleter}}{Filename Completer}
\item{\link{GFileOutputStream}}{File output streaming operations}
\item{\link{GFilterInputStream}}{Filter Input Stream}
\item{\link{GFilterOutputStream}}{Filter Output Stream}
\item{\link{GIcon}}{Interface for icons}
\item{\link{GInetAddress}}{An IPv4/IPv6 address}
\item{\link{GInetSocketAddress}}{Internet GSocketAddress}
\item{\link{GInitable}}{Failable object initialization interface}
\item{\link{GInputStream}}{Base class for implementing streaming input}
\item{\link{gio-GIOError}}{Error helper functions}
\item{\link{GIOModule}}{Loadable GIO Modules}
\item{\link{gio-GIOScheduler}}{I/O Scheduler}
\item{\link{GIOStream}}{Base class for implementing read/write streams}
\item{\link{GLoadableIcon}}{Loadable Icons}
\item{\link{GMemoryInputStream}}{Streaming input operations on memory chunks}
\item{\link{GMemoryOutputStream}}{Streaming output operations on memory chunks}
\item{\link{GMount}}{Mount management}
\item{\link{GMountOperation}}{Object used for authentication and user interaction}
\item{\link{GNetworkAddress}}{A GSocketConnectable for resolving hostnames}
\item{\link{GNetworkService}}{A GSocketConnectable for resolving SRV records}
\item{\link{GOutputStream}}{Base class for implementing streaming output}
\item{\link{GResolver}}{Asynchronous and cancellable DNS resolver}
\item{\link{GSeekable}}{Stream seeking interface}
\item{\link{GSimpleAsyncResult}}{Simple asynchronous results implementation}
\item{\link{GSocket}}{Low-level socket object}
\item{\link{GSocketAddress}}{Abstract base class representing endpoints for
socket communication}
\item{\link{GSocketClient}}{Helper for connecting to a network service}
\item{\link{GSocketConnectable}}{Interface for potential socket endpoints}
\item{\link{GSocketConnection}}{A socket connection}
\item{\link{GSocketControlMessage}}{A GSocket control message}
\item{\link{GSocketListener}}{Helper for accepting network client connections}
\item{\link{GSocketService}}{Make it easy to implement a network service}
\item{\link{GSrvTarget}}{DNS SRV record target}
\item{\link{GThemedIcon}}{Icon theming support}
\item{\link{GThreadedSocketService}}{A threaded GSocketService}
\item{\link{GVfs}}{Virtual File System}
\item{\link{GVolume}}{Volume management}
\item{\link{GVolumeMonitor}}{Volume Monitor}
}
}
\references{\url{https://developer-old.gnome.org/gio/}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{interface}
