\alias{gIoExtensionGetPriority}
\name{gIoExtensionGetPriority}
\title{gIoExtensionGetPriority}
\description{Gets the priority with which \code{extension} was registered.}
\usage{gIoExtensionGetPriority(object)}
\arguments{\item{\verb{object}}{a \code{\link{GIOExtension}}}}
\value{[integer] the priority of \code{extension}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
