\alias{gIoExtensionGetType}
\name{gIoExtensionGetType}
\title{gIoExtensionGetType}
\description{Gets the type associated with \code{extension}.}
\usage{gIoExtensionGetType(object)}
\arguments{\item{\verb{object}}{a \code{\link{GIOExtension}}}}
\value{[\code{\link{GType}}] the type of \code{extension}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
