\alias{gIoExtensionPointGetExtensions}
\name{gIoExtensionPointGetExtensions}
\title{gIoExtensionPointGetExtensions}
\description{Gets a list of all extensions that implement this extension point.
The list is sorted by priority, beginning with the highest priority.}
\usage{gIoExtensionPointGetExtensions(object)}
\arguments{\item{\verb{object}}{a \code{\link{GIOExtensionPoint}}}}
\value{[list] a \verb{list} of \code{\link{GIOExtension}}s.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
