\alias{gIoExtensionPointLookup}
\name{gIoExtensionPointLookup}
\title{gIoExtensionPointLookup}
\description{Looks up an existing extension point.}
\usage{gIoExtensionPointLookup(name)}
\arguments{\item{\verb{name}}{the name of the extension point}}
\value{[\code{\link{GIOExtensionPoint}}] the \code{\link{GIOExtensionPoint}}, or \code{NULL} if there is no
registered extension point with the given name}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
