\alias{gIoExtensionPointRegister}
\name{gIoExtensionPointRegister}
\title{gIoExtensionPointRegister}
\description{Registers an extension point.}
\usage{gIoExtensionPointRegister(name)}
\arguments{\item{\verb{name}}{The name of the extension point}}
\value{[\code{\link{GIOExtensionPoint}}] the new \code{\link{GIOExtensionPoint}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
