\alias{gIoExtensionPointSetRequiredType}
\name{gIoExtensionPointSetRequiredType}
\title{gIoExtensionPointSetRequiredType}
\description{Sets the required type for \code{extension.point} to \code{type}. 
All implementations must henceforth have this type.}
\usage{gIoExtensionPointSetRequiredType(object, type)}
\arguments{
\item{\verb{object}}{a \code{\link{GIOExtensionPoint}}}
\item{\verb{type}}{the \code{\link{GType}} to require}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
