\alias{gIoExtensionRefClass}
\name{gIoExtensionRefClass}
\title{gIoExtensionRefClass}
\description{Gets a reference to the class for the type that is 
associated with \code{extension}.}
\usage{gIoExtensionRefClass(object)}
\arguments{\item{\verb{object}}{a \code{\link{GIOExtension}}}}
\value{[GTypeClass] the \verb{GTypeClass} for the type of \code{extension}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
