\alias{gIoSchedulerCancelAllJobs}
\name{gIoSchedulerCancelAllJobs}
\title{gIoSchedulerCancelAllJobs}
\description{Cancels all cancellable I/O jobs. }
\usage{gIoSchedulerCancelAllJobs()}
\details{A job is cancellable if a \code{\link{GCancellable}} was passed into
\code{gIoSchedulerPushJob()}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
