\alias{gIOStreamClearPending}
\name{gIOStreamClearPending}
\title{gIOStreamClearPending}
\description{Clears the pending flag on \code{stream}.}
\usage{gIOStreamClearPending(object)}
\arguments{\item{\verb{object}}{a \code{\link{GIOStream}}}}
\details{Since 2.22}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
