\alias{gIOStreamGetInputStream}
\name{gIOStreamGetInputStream}
\title{gIOStreamGetInputStream}
\description{Gets the input stream for this object. This is used
for reading.}
\usage{gIOStreamGetInputStream(object)}
\arguments{\item{\verb{object}}{a \code{\link{GIOStream}}}}
\details{Since 2.22}
\value{[\code{\link{GInputStream}}] a \code{\link{GInputStream}}, Do not free.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
