\alias{gIOStreamGetOutputStream}
\name{gIOStreamGetOutputStream}
\title{gIOStreamGetOutputStream}
\description{Gets the output stream for this object. This is used for
writing.}
\usage{gIOStreamGetOutputStream(object)}
\arguments{\item{\verb{object}}{a \code{\link{GIOStream}}}}
\details{Since 2.22}
\value{[\code{\link{GOutputStream}}] a \code{\link{GOutputStream}}, Do not free.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
