\alias{gIOStreamHasPending}
\name{gIOStreamHasPending}
\title{gIOStreamHasPending}
\description{Checks if a stream has pending actions.}
\usage{gIOStreamHasPending(object)}
\arguments{\item{\verb{object}}{a \code{\link{GIOStream}}}}
\details{Since 2.22}
\value{[logical] \code{TRUE} if \code{stream} has pending actions.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
