\alias{gIOStreamIsClosed}
\name{gIOStreamIsClosed}
\title{gIOStreamIsClosed}
\description{Checks if a stream is closed.}
\usage{gIOStreamIsClosed(object)}
\arguments{\item{\verb{object}}{a \code{\link{GIOStream}}}}
\details{Since 2.22}
\value{[logical] \code{TRUE} if the stream is closed.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
