\alias{gladeProvide}
\alias{gladeRequire}
\title{Glade library loading}
\name{glade-loading}
\usage{
gladeRequire(library)
}
\arguments{
\item{library}{The name of the library to load.}
}
\description{This function might allow one to load modules into Glade.}
\author{Michael Lawrence}
\keyword{interface}
\keyword{internal}
