\alias{gladeGetWidgetName}
\name{gladeGetWidgetName}
\title{gladeGetWidgetName}
\description{Used to get the name of a widget that was generated by a GladeXML object.}
\usage{gladeGetWidgetName(widget)}
\arguments{\item{\code{widget}}{[\code{\link{GtkWidget}}]  the widget}}
\value{[char]  the name of the widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
