\alias{gMemoryInputStreamAddData}
\name{gMemoryInputStreamAddData}
\title{gMemoryInputStreamAddData}
\description{Appends \code{data} to data that can be read from the input stream}
\usage{gMemoryInputStreamAddData(object, data)}
\arguments{
\item{\verb{object}}{a \code{\link{GMemoryInputStream}}}
\item{\verb{data}}{input data}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
