\alias{gMemoryInputStreamNew}
\name{gMemoryInputStreamNew}
\title{gMemoryInputStreamNew}
\description{Creates a new empty \code{\link{GMemoryInputStream}}.}
\usage{gMemoryInputStreamNew()}
\value{[\code{\link{GInputStream}}] a new \code{\link{GInputStream}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
