\alias{gMemoryInputStreamNewFromData}
\name{gMemoryInputStreamNewFromData}
\title{gMemoryInputStreamNewFromData}
\description{Creates a new \code{\link{GMemoryInputStream}} with data in memory of a given size.}
\usage{gMemoryInputStreamNewFromData(data)}
\arguments{\item{\verb{data}}{input data}}
\value{[\code{\link{GInputStream}}] new \code{\link{GInputStream}} read from \code{data} of \code{len} bytes.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
