\alias{gMemoryOutputStreamGetDataSize}
\name{gMemoryOutputStreamGetDataSize}
\title{gMemoryOutputStreamGetDataSize}
\description{Returns the number of bytes from the start up
to including the last byte written in the stream
that has not been truncated away.}
\usage{gMemoryOutputStreamGetDataSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GMemoryOutputStream}}}}
\details{Since 2.18}
\value{[numeric] the number of bytes written to the stream}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
