\alias{gMountCanEject}
\name{gMountCanEject}
\title{gMountCanEject}
\description{Checks if \code{mount} can be eject.}
\usage{gMountCanEject(object)}
\arguments{\item{\verb{object}}{a \code{\link{GMount}}.}}
\value{[logical] \code{TRUE} if the \code{mount} can be ejected.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
