\alias{gMountCanUnmount}
\name{gMountCanUnmount}
\title{gMountCanUnmount}
\description{Checks if \code{mount} can be mounted.}
\usage{gMountCanUnmount(object)}
\arguments{\item{\verb{object}}{a \code{\link{GMount}}.}}
\value{[logical] \code{TRUE} if the \code{mount} can be unmounted.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
