\alias{gMountGetIcon}
\name{gMountGetIcon}
\title{gMountGetIcon}
\description{Gets the icon for \code{mount}.}
\usage{gMountGetIcon(object)}
\arguments{\item{\verb{object}}{a \code{\link{GMount}}.}}
\value{[\code{\link{GIcon}}] a \code{\link{GIcon}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
