\alias{gMountGetName}
\name{gMountGetName}
\title{gMountGetName}
\description{Gets the name of \code{mount}.}
\usage{gMountGetName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GMount}}.}}
\value{[char] the name for the given \code{mount}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
