\alias{gMountGetRoot}
\name{gMountGetRoot}
\title{gMountGetRoot}
\description{Gets the root directory on \code{mount}.}
\usage{gMountGetRoot(object)}
\arguments{\item{\verb{object}}{a \code{\link{GMount}}.}}
\value{[\code{\link{GFile}}] a \code{\link{GFile}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
