\alias{gMountGetVolume}
\name{gMountGetVolume}
\title{gMountGetVolume}
\description{Gets the volume for the \code{mount}.}
\usage{gMountGetVolume(object)}
\arguments{\item{\verb{object}}{a \code{\link{GMount}}.}}
\value{[\code{\link{GVolume}}] a \code{\link{GVolume}} or \code{NULL} if \code{mount} is not associated with a volume.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
