\alias{gMountOperationGetAnonymous}
\name{gMountOperationGetAnonymous}
\title{gMountOperationGetAnonymous}
\description{Check to see whether the mount operation is being used 
for an anonymous user.}
\usage{gMountOperationGetAnonymous(object)}
\arguments{\item{\verb{object}}{a \code{\link{GMountOperation}}.}}
\value{[logical] \code{TRUE} if mount operation is anonymous.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
