\alias{gMountOperationGetChoice}
\name{gMountOperationGetChoice}
\title{gMountOperationGetChoice}
\description{Gets a choice from the mount operation.}
\usage{gMountOperationGetChoice(object)}
\arguments{\item{\verb{object}}{a \code{\link{GMountOperation}}.}}
\value{[integer] an integer containing an index of the user's choice from
the choice's list, or \code{0}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
