\alias{gMountOperationGetDomain}
\name{gMountOperationGetDomain}
\title{gMountOperationGetDomain}
\description{Gets the domain of the mount operation.}
\usage{gMountOperationGetDomain(object)}
\arguments{\item{\verb{object}}{a \code{\link{GMountOperation}}.}}
\value{[char] a string set to the domain.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
