\alias{gMountOperationGetPassword}
\name{gMountOperationGetPassword}
\title{gMountOperationGetPassword}
\description{Gets a password from the mount operation.}
\usage{gMountOperationGetPassword(object)}
\arguments{\item{\verb{object}}{a \code{\link{GMountOperation}}.}}
\value{[char] a string containing the password within \code{op}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
