\alias{gMountOperationGetPasswordSave}
\name{gMountOperationGetPasswordSave}
\title{gMountOperationGetPasswordSave}
\description{Gets the state of saving passwords for the mount operation.}
\usage{gMountOperationGetPasswordSave(object)}
\arguments{\item{\verb{object}}{a \code{\link{GMountOperation}}.}}
\value{[\code{\link{GPasswordSave}}] a \code{\link{GPasswordSave}} flag.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
