\alias{gMountOperationGetUsername}
\name{gMountOperationGetUsername}
\title{gMountOperationGetUsername}
\description{Get the user name from the mount operation.}
\usage{gMountOperationGetUsername(object)}
\arguments{\item{\verb{object}}{a \code{\link{GMountOperation}}.}}
\value{[char] a string containing the user name.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
