\alias{gMountOperationReply}
\name{gMountOperationReply}
\title{gMountOperationReply}
\description{Emits the \code{\link{gMountOperationReply}} signal.}
\usage{gMountOperationReply(object, result)}
\arguments{
\item{\verb{object}}{a \code{\link{GMountOperation}}}
\item{\verb{result}}{a \code{\link{GMountOperationResult}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
