\alias{gMountOperationSetAnonymous}
\name{gMountOperationSetAnonymous}
\title{gMountOperationSetAnonymous}
\description{Sets the mount operation to use an anonymous user if \code{anonymous} is \code{TRUE}.}
\usage{gMountOperationSetAnonymous(object, anonymous)}
\arguments{
\item{\verb{object}}{a \code{\link{GMountOperation}}.}
\item{\verb{anonymous}}{boolean value.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
