\alias{gMountOperationSetChoice}
\name{gMountOperationSetChoice}
\title{gMountOperationSetChoice}
\description{Sets a default choice for the mount operation.}
\usage{gMountOperationSetChoice(object, choice)}
\arguments{
\item{\verb{object}}{a \code{\link{GMountOperation}}.}
\item{\verb{choice}}{an integer.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
