\alias{gMountOperationSetDomain}
\name{gMountOperationSetDomain}
\title{gMountOperationSetDomain}
\description{Sets the mount operation's domain.}
\usage{gMountOperationSetDomain(object, domain)}
\arguments{
\item{\verb{object}}{a \code{\link{GMountOperation}}.}
\item{\verb{domain}}{the domain to set.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
