\alias{gMountOperationSetPassword}
\name{gMountOperationSetPassword}
\title{gMountOperationSetPassword}
\description{Sets the mount operation's password to \code{password}.}
\usage{gMountOperationSetPassword(object, password)}
\arguments{
\item{\verb{object}}{a \code{\link{GMountOperation}}.}
\item{\verb{password}}{password to set.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
