\alias{gMountOperationSetPasswordSave}
\name{gMountOperationSetPasswordSave}
\title{gMountOperationSetPasswordSave}
\description{Sets the state of saving passwords for the mount operation.}
\usage{gMountOperationSetPasswordSave(object, save)}
\arguments{
\item{\verb{object}}{a \code{\link{GMountOperation}}.}
\item{\verb{save}}{a set of \code{\link{GPasswordSave}} flags.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
