\alias{gMountOperationSetUsername}
\name{gMountOperationSetUsername}
\title{gMountOperationSetUsername}
\description{Sets the user name within \code{op} to \code{username}.}
\usage{gMountOperationSetUsername(object, username)}
\arguments{
\item{\verb{object}}{a \code{\link{GMountOperation}}.}
\item{\verb{username}}{input username.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
