\alias{gNetworkAddressGetHostname}
\name{gNetworkAddressGetHostname}
\title{gNetworkAddressGetHostname}
\description{Gets \code{addr}'s hostname. This might be either UTF-8 or ASCII-encoded,
depending on what \code{addr} was created with.}
\usage{gNetworkAddressGetHostname(object)}
\arguments{\item{\verb{object}}{a \code{\link{GNetworkAddress}}}}
\details{Since 2.22}
\value{[character] \code{addr}'s hostname}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
