\alias{gNetworkAddressGetPort}
\name{gNetworkAddressGetPort}
\title{gNetworkAddressGetPort}
\description{Gets \code{addr}'s port number}
\usage{gNetworkAddressGetPort(object)}
\arguments{\item{\verb{object}}{a \code{\link{GNetworkAddress}}}}
\details{Since 2.22}
\value{[integer] \code{addr}'s port (which may be 0)}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
