\alias{gNetworkAddressNew}
\name{gNetworkAddressNew}
\title{gNetworkAddressNew}
\description{Creates a new \code{\link{GSocketConnectable}} for connecting to the given
\code{hostname} and \code{port}.}
\usage{gNetworkAddressNew(hostname, port)}
\arguments{
\item{\verb{hostname}}{the hostname}
\item{\verb{port}}{the port}
}
\details{Since 2.22}
\value{[\code{\link{GSocketConnectable}}] the new \code{\link{GNetworkAddress}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
