\alias{gNetworkServiceGetDomain}
\name{gNetworkServiceGetDomain}
\title{gNetworkServiceGetDomain}
\description{Gets the domain that \code{srv} serves. This might be either UTF-8 or
ASCII-encoded, depending on what \code{srv} was created with.}
\usage{gNetworkServiceGetDomain(object)}
\arguments{\item{\verb{object}}{a \code{\link{GNetworkService}}}}
\details{Since 2.22}
\value{[character] \code{srv}'s domain name}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
