\alias{gNetworkServiceGetProtocol}
\name{gNetworkServiceGetProtocol}
\title{gNetworkServiceGetProtocol}
\description{Gets \code{srv}'s protocol name (eg, "tcp").}
\usage{gNetworkServiceGetProtocol(object)}
\arguments{\item{\verb{object}}{a \code{\link{GNetworkService}}}}
\details{Since 2.22}
\value{[character] \code{srv}'s protocol name}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
