\alias{gNetworkServiceGetService}
\name{gNetworkServiceGetService}
\title{gNetworkServiceGetService}
\description{Gets \code{srv}'s service name (eg, "ldap").}
\usage{gNetworkServiceGetService(object)}
\arguments{\item{\verb{object}}{a \code{\link{GNetworkService}}}}
\details{Since 2.22}
\value{[character] \code{srv}'s service name}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
