\alias{gOutputStreamClearPending}
\name{gOutputStreamClearPending}
\title{gOutputStreamClearPending}
\description{Clears the pending flag on \code{stream}.}
\usage{gOutputStreamClearPending(object)}
\arguments{\item{\verb{object}}{output stream}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
