\alias{gOutputStreamHasPending}
\name{gOutputStreamHasPending}
\title{gOutputStreamHasPending}
\description{Checks if an ouput stream has pending actions.}
\usage{gOutputStreamHasPending(object)}
\arguments{\item{\verb{object}}{a \code{\link{GOutputStream}}.}}
\value{[logical] \code{TRUE} if \code{stream} has pending actions.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
