\alias{gOutputStreamIsClosed}
\name{gOutputStreamIsClosed}
\title{gOutputStreamIsClosed}
\description{Checks if an output stream has already been closed.}
\usage{gOutputStreamIsClosed(object)}
\arguments{\item{\verb{object}}{a \code{\link{GOutputStream}}.}}
\value{[logical] \code{TRUE} if \code{stream} is closed. \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
