\alias{gResolverFreeTargets}
\name{gResolverFreeTargets}
\title{gResolverFreeTargets}
\description{Frees \code{targets} (which should be the return value from
\code{\link{gResolverLookupService}} or \code{\link{gResolverLookupServiceFinish}}).
(This is a convenience method; you can also simply free the
results by hand.)}
\usage{gResolverFreeTargets(targets)}
\arguments{\item{\verb{targets}}{a \verb{list} of \code{\link{GSrvTarget}}}}
\details{Since 2.22}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
