\alias{gSeekableCanSeek}
\name{gSeekableCanSeek}
\title{gSeekableCanSeek}
\description{Tests if the stream supports the \verb{GSeekableIface}.}
\usage{gSeekableCanSeek(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSeekable}}.}}
\value{[logical] \code{TRUE} if \code{seekable} can be seeked. \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
