\alias{gSeekableCanTruncate}
\name{gSeekableCanTruncate}
\title{gSeekableCanTruncate}
\description{Tests if the stream can be truncated.}
\usage{gSeekableCanTruncate(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSeekable}}.}}
\value{[logical] \code{TRUE} if the stream can be truncated, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
