\alias{gSeekableTell}
\name{gSeekableTell}
\title{gSeekableTell}
\description{Tells the current position within the stream.}
\usage{gSeekableTell(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSeekable}}.}}
\value{[numeric] the offset from the beginning of the buffer.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
