\alias{gSimpleAsyncResultCompleteInIdle}
\name{gSimpleAsyncResultCompleteInIdle}
\title{gSimpleAsyncResultCompleteInIdle}
\description{Completes an asynchronous function in an idle handler in the thread-default main
loop of the thread that \code{simple} was initially created in.}
\usage{gSimpleAsyncResultCompleteInIdle(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSimpleAsyncResult}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
