\alias{gSimpleAsyncResultGetOpResGboolean}
\name{gSimpleAsyncResultGetOpResGboolean}
\title{gSimpleAsyncResultGetOpResGboolean}
\description{Gets the operation result boolean from within the asynchronous result.}
\usage{gSimpleAsyncResultGetOpResGboolean(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSimpleAsyncResult}}.}}
\value{[logical] \code{TRUE} if the operation's result was \code{TRUE}, \code{FALSE}
if the operation's result was \code{FALSE}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
