\alias{gSimpleAsyncResultGetOpResGpointer}
\name{gSimpleAsyncResultGetOpResGpointer}
\title{gSimpleAsyncResultGetOpResGpointer}
\description{Gets a pointer result as returned by the asynchronous function.}
\usage{gSimpleAsyncResultGetOpResGpointer(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSimpleAsyncResult}}.}}
\value{[R object] a pointer from the result.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
