\alias{gSimpleAsyncResultGetOpResGssize}
\name{gSimpleAsyncResultGetOpResGssize}
\title{gSimpleAsyncResultGetOpResGssize}
\description{Gets a gssize from the asynchronous result.}
\usage{gSimpleAsyncResultGetOpResGssize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSimpleAsyncResult}}.}}
\value{[integer] a gssize returned from the asynchronous function.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
