\alias{gSimpleAsyncResultGetSourceTag}
\name{gSimpleAsyncResultGetSourceTag}
\title{gSimpleAsyncResultGetSourceTag}
\description{Gets the source tag for the \code{\link{GSimpleAsyncResult}}.}
\usage{gSimpleAsyncResultGetSourceTag(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSimpleAsyncResult}}.}}
\value{[R object] a \verb{R object} to the source object for the \code{\link{GSimpleAsyncResult}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
