\alias{gSimpleAsyncResultSetFromError}
\name{gSimpleAsyncResultSetFromError}
\title{gSimpleAsyncResultSetFromError}
\description{Sets the result from a \code{\link{GError}}.}
\usage{gSimpleAsyncResultSetFromError(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSimpleAsyncResult}}.}}
\value{
A list containing the following elements:
\item{\verb{error}}{\code{\link{GError}}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
