\alias{gSimpleAsyncResultSetHandleCancellation}
\name{gSimpleAsyncResultSetHandleCancellation}
\title{gSimpleAsyncResultSetHandleCancellation}
\description{Sets whether to handle cancellation within the asynchronous operation.}
\usage{gSimpleAsyncResultSetHandleCancellation(object, handle.cancellation)}
\arguments{
\item{\verb{object}}{a \code{\link{GSimpleAsyncResult}}.}
\item{\verb{handle.cancellation}}{a \verb{logical}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
