\alias{gSimpleAsyncResultSetOpResGboolean}
\name{gSimpleAsyncResultSetOpResGboolean}
\title{gSimpleAsyncResultSetOpResGboolean}
\description{Sets the operation result to a boolean within the asynchronous result.}
\usage{gSimpleAsyncResultSetOpResGboolean(object, op.res)}
\arguments{
\item{\verb{object}}{a \code{\link{GSimpleAsyncResult}}.}
\item{\verb{op.res}}{a \verb{logical}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
