\alias{gSimpleAsyncResultSetOpResGpointer}
\name{gSimpleAsyncResultSetOpResGpointer}
\title{gSimpleAsyncResultSetOpResGpointer}
\description{Sets the operation result within the asynchronous result to a pointer.}
\usage{gSimpleAsyncResultSetOpResGpointer(object, op.res)}
\arguments{
\item{\verb{object}}{a \code{\link{GSimpleAsyncResult}}.}
\item{\verb{op.res}}{a pointer result from an asynchronous function.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
