\alias{gSimpleAsyncResultSetOpResGssize}
\name{gSimpleAsyncResultSetOpResGssize}
\title{gSimpleAsyncResultSetOpResGssize}
\description{Sets the operation result within the asynchronous result to 
the given \code{op.res}.}
\usage{gSimpleAsyncResultSetOpResGssize(object, op.res)}
\arguments{
\item{\verb{object}}{a \code{\link{GSimpleAsyncResult}}.}
\item{\verb{op.res}}{a \verb{integer}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
