\alias{gSocketAddressGetFamily}
\name{gSocketAddressGetFamily}
\title{gSocketAddressGetFamily}
\description{Gets the socket family type of \code{address}.}
\usage{gSocketAddressGetFamily(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSocketAddress}}}}
\details{Since 2.22}
\value{[\code{\link{GSocketFamily}}] the socket family type of \code{address}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
