\alias{gSocketClientGetFamily}
\name{gSocketClientGetFamily}
\title{gSocketClientGetFamily}
\description{Gets the socket family of the socket client.}
\usage{gSocketClientGetFamily(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSocketClient}}.}}
\details{See \code{\link{gSocketClientSetFamily}} for details.
  Since 2.22}
\value{[\code{\link{GSocketFamily}}] a \code{\link{GSocketFamily}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
