\alias{gSocketClientGetLocalAddress}
\name{gSocketClientGetLocalAddress}
\title{gSocketClientGetLocalAddress}
\description{Gets the local the socket client.}
\usage{gSocketClientGetLocalAddress(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSocketClient}}.}}
\details{See \code{\link{gSocketClientSetLocalAddress}} for details.
  Since 2.22}
\value{[\code{\link{GSocketAddress}}] a \verb{GSocketAddres} or \code{NULL}. don't free}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
