\alias{gSocketClientGetProtocol}
\name{gSocketClientGetProtocol}
\title{gSocketClientGetProtocol}
\description{Gets the protocol name type of the socket client.}
\usage{gSocketClientGetProtocol(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSocketClient}}}}
\details{See \code{\link{gSocketClientSetProtocol}} for details.
  Since 2.22}
\value{[\code{\link{GSocketProtocol}}] a \code{\link{GSocketProtocol}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
