\alias{gSocketClientGetSocketType}
\name{gSocketClientGetSocketType}
\title{gSocketClientGetSocketType}
\description{Gets the socket type of the socket client.}
\usage{gSocketClientGetSocketType(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSocketClient}}.}}
\details{See \code{\link{gSocketClientSetSocketType}} for details.
  Since 2.22}
\value{[\code{\link{GSocketType}}] a \code{\link{GSocketFamily}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
