\alias{gSocketClientNew}
\name{gSocketClientNew}
\title{gSocketClientNew}
\description{Creates a new \code{\link{GSocketClient}} with the default options.}
\usage{gSocketClientNew()}
\details{Since 2.22}
\value{[\code{\link{GSocketClient}}] a \code{\link{GSocketClient}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
