\alias{gSocketConnectableEnumerate}
\name{gSocketConnectableEnumerate}
\title{gSocketConnectableEnumerate}
\description{Creates a \code{\link{GSocketAddressEnumerator}} for \code{connectable}.}
\usage{gSocketConnectableEnumerate(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSocketConnectable}}}}
\details{Since 2.22}
\value{[\code{\link{GSocketAddressEnumerator}}] a new \code{\link{GSocketAddressEnumerator}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
