\alias{gSocketConnectionFactoryCreateConnection}
\name{gSocketConnectionFactoryCreateConnection}
\title{gSocketConnectionFactoryCreateConnection}
\description{Creates a \code{\link{GSocketConnection}} subclass of the right type for
\code{socket}.}
\usage{gSocketConnectionFactoryCreateConnection(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSocket}}}}
\details{Since 2.22}
\value{[\code{\link{GSocketConnection}}] a \code{\link{GSocketConnection}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
