\alias{gSocketControlMessageGetLevel}
\name{gSocketControlMessageGetLevel}
\title{gSocketControlMessageGetLevel}
\description{Returns the "level" (i.e. the originating protocol) of the control message.
This is often SOL_SOCKET.}
\usage{gSocketControlMessageGetLevel(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSocketControlMessage}}}}
\details{Since 2.22}
\value{[integer] an integer describing the level}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
