\alias{gSocketControlMessageGetSize}
\name{gSocketControlMessageGetSize}
\title{gSocketControlMessageGetSize}
\description{Returns the space required for the control message, not including
headers or alignment.}
\usage{gSocketControlMessageGetSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSocketControlMessage}}}}
\details{Since 2.22}
\value{[numeric] The number of bytes required.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
