\alias{gSocketGetBlocking}
\name{gSocketGetBlocking}
\title{gSocketGetBlocking}
\description{Gets the blocking mode of the socket. For details on blocking I/O,
see \code{\link{gSocketSetBlocking}}.}
\usage{gSocketGetBlocking(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSocket}}.}}
\details{Since 2.22}
\value{[logical] \code{TRUE} if blocking I/O is used, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
