\alias{gSocketGetFamily}
\name{gSocketGetFamily}
\title{gSocketGetFamily}
\description{Gets the socket family of the socket.}
\usage{gSocketGetFamily(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSocket}}.}}
\details{Since 2.22}
\value{[\code{\link{GSocketFamily}}] a \code{\link{GSocketFamily}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
