\alias{gSocketGetKeepalive}
\name{gSocketGetKeepalive}
\title{gSocketGetKeepalive}
\description{Gets the keepalive mode of the socket. For details on this,
see \code{\link{gSocketSetKeepalive}}.}
\usage{gSocketGetKeepalive(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSocket}}.}}
\details{Since 2.22}
\value{[logical] \code{TRUE} if keepalive is active, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
