\alias{gSocketGetSocketType}
\name{gSocketGetSocketType}
\title{gSocketGetSocketType}
\description{Gets the socket type of the socket.}
\usage{gSocketGetSocketType(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSocket}}.}}
\details{Since 2.22}
\value{[\code{\link{GSocketType}}] a \code{\link{GSocketType}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
