\alias{gSocketIsClosed}
\name{gSocketIsClosed}
\title{gSocketIsClosed}
\description{Checks whether a socket is closed.}
\usage{gSocketIsClosed(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSocket}}}}
\details{Since 2.22}
\value{[logical] \code{TRUE} if socket is closed, \code{FALSE} otherwise}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
