\alias{gSocketIsConnected}
\name{gSocketIsConnected}
\title{gSocketIsConnected}
\description{Check whether the socket is connected. This is only useful for
connection-oriented sockets.}
\usage{gSocketIsConnected(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSocket}}.}}
\details{Since 2.22}
\value{[logical] \code{TRUE} if socket is connected, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
