\alias{gSocketListenerClose}
\name{gSocketListenerClose}
\title{gSocketListenerClose}
\description{Closes all the sockets in the listener.}
\usage{gSocketListenerClose(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSocketListener}}}}
\details{Since 2.22}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
