\alias{gSocketListenerNew}
\name{gSocketListenerNew}
\title{gSocketListenerNew}
\description{Creates a new \code{\link{GSocketListener}} with no sockets to listen for.
New listeners can be added with e.g. \code{\link{gSocketListenerAddAddress}}
or \code{\link{gSocketListenerAddInetPort}}.}
\usage{gSocketListenerNew()}
\details{Since 2.22}
\value{[\code{\link{GSocketListener}}] a new \code{\link{GSocketListener}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
