\alias{gSocketListenerSetBacklog}
\name{gSocketListenerSetBacklog}
\title{gSocketListenerSetBacklog}
\description{Sets the listen backlog on the sockets in the listener.}
\usage{gSocketListenerSetBacklog(object, listen.backlog)}
\arguments{
\item{\verb{object}}{a \code{\link{GSocketListener}}}
\item{\verb{listen.backlog}}{an integer}
}
\details{See \code{\link{gSocketSetListenBacklog}} for details
  Since 2.22}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
