\alias{gSocketServiceNew}
\name{gSocketServiceNew}
\title{gSocketServiceNew}
\description{Creates a new \code{\link{GSocketService}} with no sockets to listen for.
New listeners can be added with e.g. \code{\link{gSocketListenerAddAddress}}
or \code{\link{gSocketListenerAddInetPort}}.}
\usage{gSocketServiceNew()}
\details{Since 2.22}
\value{[\code{\link{GSocketService}}] a new \code{\link{GSocketService}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
