\alias{gSrvTargetCopy}
\name{gSrvTargetCopy}
\title{gSrvTargetCopy}
\description{Copies \code{target}}
\usage{gSrvTargetCopy(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSrvTarget}}}}
\details{Since 2.22}
\value{[\code{\link{GSrvTarget}}] a copy of \code{target}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
