\alias{gSrvTargetGetPort}
\name{gSrvTargetGetPort}
\title{gSrvTargetGetPort}
\description{Gets \code{target}'s port}
\usage{gSrvTargetGetPort(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSrvTarget}}}}
\details{Since 2.22}
\value{[integer] \code{target}'s port}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
