\alias{gSrvTargetGetPriority}
\name{gSrvTargetGetPriority}
\title{gSrvTargetGetPriority}
\description{Gets \code{target}'s priority. You should not need to look at this;
\code{\link{GResolver}} already sorts the targets according to the algorithm in
RFC 2782.}
\usage{gSrvTargetGetPriority(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSrvTarget}}}}
\details{Since 2.22}
\value{[integer] \code{target}'s priority}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
