\alias{gSrvTargetGetWeight}
\name{gSrvTargetGetWeight}
\title{gSrvTargetGetWeight}
\description{Gets \code{target}'s weight. You should not need to look at this;
\code{\link{GResolver}} already sorts the targets according to the algorithm in
RFC 2782.}
\usage{gSrvTargetGetWeight(object)}
\arguments{\item{\verb{object}}{a \code{\link{GSrvTarget}}}}
\details{Since 2.22}
\value{[integer] \code{target}'s weight}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
