\alias{gSrvTargetListSort}
\name{gSrvTargetListSort}
\title{gSrvTargetListSort}
\description{Sorts \code{targets} in place according to the algorithm in RFC 2782.}
\usage{gSrvTargetListSort(targets)}
\arguments{\item{\verb{targets}}{a \verb{list} of \code{\link{GSrvTarget}}}}
\details{Since 2.22}
\value{[list] the head of the sorted list.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
