\alias{gTcpConnectionGetGracefulDisconnect}
\name{gTcpConnectionGetGracefulDisconnect}
\title{gTcpConnectionGetGracefulDisconnect}
\description{Checks if graceful disconnects are used. See
\code{\link{gTcpConnectionSetGracefulDisconnect}}.}
\usage{gTcpConnectionGetGracefulDisconnect(object)}
\arguments{\item{\verb{object}}{a \code{\link{GTcpConnection}}}}
\details{Since 2.22}
\value{[logical] \code{TRUE} if graceful disconnect is used on close, \code{FALSE} otherwise}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
