\alias{gThemedIconAppendName}
\name{gThemedIconAppendName}
\title{gThemedIconAppendName}
\description{Append a name to the list of icons from within \code{icon}.}
\usage{gThemedIconAppendName(object, iconname)}
\arguments{
\item{\verb{object}}{a \code{\link{GThemedIcon}}}
\item{\verb{iconname}}{name of icon to append to list of icons from within \code{icon}.}
}
\details{\strong{PLEASE NOTE:} Note that doing so invalidates the hash computed by prior calls
to \code{\link{gIconHash}}.}
\note{Note that doing so invalidates the hash computed by prior calls
to \code{\link{gIconHash}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
