\alias{gThemedIconGetNames}
\name{gThemedIconGetNames}
\title{gThemedIconGetNames}
\description{Gets the names of icons from within \code{icon}.}
\usage{gThemedIconGetNames(object)}
\arguments{\item{\verb{object}}{a \code{\link{GThemedIcon}}.}}
\value{[character] a list of icon names.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
