\alias{gThemedIconNew}
\name{gThemedIconNew}
\title{gThemedIconNew}
\description{Creates a new themed icon for \code{iconname}.}
\usage{gThemedIconNew(iconname = NULL)}
\arguments{\item{\verb{iconname}}{a string containing an icon name.}}
\value{[\code{\link{GIcon}}] a new \code{\link{GThemedIcon}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
