\alias{gThemedIconNewFromNames}
\name{gThemedIconNewFromNames}
\title{gThemedIconNewFromNames}
\description{Creates a new themed icon for \code{iconnames}.}
\usage{gThemedIconNewFromNames(iconnames, len)}
\arguments{
\item{\verb{iconnames}}{a list of strings containing icon names.}
\item{\verb{len}}{the length of the \code{iconnames} list, or -1 if \code{iconnames} is
\code{NULL}-terminated}
}
\value{[\code{\link{GIcon}}] a new \code{\link{GThemedIcon}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
