\name{GTimeVal}
\alias{GTimeVal}
\title{The GTimeVal structure}
\description{Represents a precise time, with seconds and microseconds.}
\details{
This is a transparent type, represented by a list with the following fields:
\describe{
\item{tv_secs}{seconds}
\item{tv_usecs}{microseconds}
}
}
\author{Michael Lawrence}
\keyword{interface}
\keyword{internal}
