\alias{gtk-Testing}
\name{gtk-Testing}
\title{Testing}
\description{Utilities for testing GTK+ applications}
\section{Methods and Functions}{
\code{\link{gtkTestFindLabel}(widget, label.pattern)}\cr
\code{\link{gtkTestFindSibling}(base.widget, widget.type)}\cr
\code{\link{gtkTestFindWidget}(widget, label.pattern, widget.type)}\cr
\code{\link{gtkTestListAllTypes}()}\cr
\code{\link{gtkTestRegisterAllTypes}()}\cr
\code{\link{gtkTestSliderGetValue}(widget)}\cr
\code{\link{gtkTestSliderSetPerc}(widget, percentage)}\cr
\code{\link{gtkTestSpinButtonClick}(spinner, button, upwards)}\cr
\code{\link{gtkTestTextGet}(widget)}\cr
\code{\link{gtkTestTextSet}(widget, string)}\cr
\code{\link{gtkTestWidgetClick}(widget, button, modifiers)}\cr
\code{\link{gtkTestWidgetSendKey}(widget, keyval, modifiers)}\cr
}
\references{\url{https://developer-old.gnome.org/gtk2/stable/gtk2-Testing.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
