\alias{gtkAboutDialogGetArtists}
\name{gtkAboutDialogGetArtists}
\title{gtkAboutDialogGetArtists}
\description{Returns the string which are displayed in the artists tab
of the secondary credits dialog.}
\usage{gtkAboutDialogGetArtists(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}}
\details{Since 2.6}
\value{[character] A string list containing
the artists.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
