\alias{gtkAboutDialogGetAuthors}
\name{gtkAboutDialogGetAuthors}
\title{gtkAboutDialogGetAuthors}
\description{Returns the string which are displayed in the authors tab
of the secondary credits dialog.}
\usage{gtkAboutDialogGetAuthors(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}}
\details{Since 2.6}
\value{[character] A string list containing
the authors.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
