\alias{gtkAboutDialogGetComments}
\name{gtkAboutDialogGetComments}
\title{gtkAboutDialogGetComments}
\description{Returns the comments string.}
\usage{gtkAboutDialogGetComments(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkAboutDialog}}]  a \verb{\link{GtkAboutDialog}}}}
\details{  Since  2.6}
\value{[character]  The comments.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
