\alias{gtkAboutDialogGetLogo}
\name{gtkAboutDialogGetLogo}
\title{gtkAboutDialogGetLogo}
\description{Returns the pixbuf displayed as logo in the about dialog.}
\usage{gtkAboutDialogGetLogo(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkAboutDialog}}]  a \verb{\link{GtkAboutDialog}}}}
\details{  Since  2.6}
\value{[\verb{\link{GdkPixbuf}}]  the pixbuf displayed as logo. If you want to keep a reference
  to it.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
