\alias{gtkAboutDialogGetLogoIconName}
\name{gtkAboutDialogGetLogoIconName}
\title{gtkAboutDialogGetLogoIconName}
\description{Returns the icon name displayed as logo in the about dialog.}
\usage{gtkAboutDialogGetLogoIconName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}}
\details{Since 2.6}
\value{[character] the icon name displayed as logo. If you want to keep a reference
to it, you have to call \code{gStrdup()} on it.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
