\alias{gtkAboutDialogGetName}
\name{gtkAboutDialogGetName}
\title{gtkAboutDialogGetName}
\description{Returns the program name displayed in the about dialog.}
\usage{gtkAboutDialogGetName(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkAboutDialog}}]  a \code{\link{GtkAboutDialog}}}}
\details{  Since  2.6}
\value{[character]  The program name.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
