\alias{gtkAboutDialogGetProgramName}
\name{gtkAboutDialogGetProgramName}
\title{gtkAboutDialogGetProgramName}
\description{Returns the program name displayed in the about dialog.}
\usage{gtkAboutDialogGetProgramName(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkAboutDialog}}]  a \verb{\link{GtkAboutDialog}}}}
\details{  Since  2.12}
\value{[character]  The program name.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
