\alias{gtkAboutDialogGetTranslatorCredits}
\name{gtkAboutDialogGetTranslatorCredits}
\title{gtkAboutDialogGetTranslatorCredits}
\description{Returns the translator credits string which is displayed
in the translators tab of the secondary credits dialog.}
\usage{gtkAboutDialogGetTranslatorCredits(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}}
\details{Since 2.6}
\value{[character] The translator credits string.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
