\alias{gtkAboutDialogGetVersion}
\name{gtkAboutDialogGetVersion}
\title{gtkAboutDialogGetVersion}
\description{Returns the version string.}
\usage{gtkAboutDialogGetVersion(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkAboutDialog}}]  a \verb{\link{GtkAboutDialog}}}}
\details{  Since  2.6}
\value{[character]  The version string.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
