\alias{gtkAboutDialogGetWebsite}
\name{gtkAboutDialogGetWebsite}
\title{gtkAboutDialogGetWebsite}
\description{Returns the website URL.}
\usage{gtkAboutDialogGetWebsite(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkAboutDialog}}]  a \code{\link{GtkAboutDialog}}}}
\details{  Since  2.6}
\value{[character]  The website URL.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
