\alias{gtkAboutDialogGetWrapLicense}
\name{gtkAboutDialogGetWrapLicense}
\title{gtkAboutDialogGetWrapLicense}
\description{Returns whether the license text in \code{about} is
automatically wrapped.}
\usage{gtkAboutDialogGetWrapLicense(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}}
\details{Since 2.8}
\value{[logical] \code{TRUE} if the license text is wrapped}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
