\alias{gtkAboutDialogSetArtists}
\name{gtkAboutDialogSetArtists}
\title{gtkAboutDialogSetArtists}
\description{Sets the strings which are displayed in the artists tab
of the secondary credits dialog.}
\usage{gtkAboutDialogSetArtists(object, artists)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAboutDialog}}]  a \code{\link{GtkAboutDialog}}}
\item{\code{artists}}{[character]  a list of strings }
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
