\alias{gtkAboutDialogSetAuthors}
\name{gtkAboutDialogSetAuthors}
\title{gtkAboutDialogSetAuthors}
\description{Sets the strings which are displayed in the authors tab
of the secondary credits dialog.}
\usage{gtkAboutDialogSetAuthors(object, authors)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAboutDialog}}]  a \code{\link{GtkAboutDialog}}}
\item{\code{authors}}{[character]  a list of strings }
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
