\alias{gtkAboutDialogSetComments}
\name{gtkAboutDialogSetComments}
\title{gtkAboutDialogSetComments}
\description{Sets the comments string to display in the about 
dialog. This should be a short string of one or
two lines.}
\usage{gtkAboutDialogSetComments(object, comments = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAboutDialog}}]  a \code{\link{GtkAboutDialog}}}
\item{\code{comments}}{[character]  a comments string}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
