\alias{gtkAboutDialogSetDocumenters}
\name{gtkAboutDialogSetDocumenters}
\title{gtkAboutDialogSetDocumenters}
\description{Sets the strings which are displayed in the documenters tab
of the secondary credits dialog.}
\usage{gtkAboutDialogSetDocumenters(object, documenters)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAboutDialog}}]  a \code{\link{GtkAboutDialog}}}
\item{\code{documenters}}{[character]  a list of strings }
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
