\alias{gtkAboutDialogSetLogo}
\name{gtkAboutDialogSetLogo}
\title{gtkAboutDialogSetLogo}
\description{Sets the pixbuf to be displayed as logo in 
the about dialog. If it is \code{NULL}, the default
window icon set with \code{\link{gtkWindowSetDefaultIcon}}
will be used.}
\usage{gtkAboutDialogSetLogo(object, logo = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkAboutDialog}}]  a \verb{\link{GtkAboutDialog}}}
\item{\verb{logo}}{[\verb{\link{GdkPixbuf}}]  a \verb{\link{GdkPixbuf}}, or \code{NULL}}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
