\alias{gtkAboutDialogSetLogoIconName}
\name{gtkAboutDialogSetLogoIconName}
\title{gtkAboutDialogSetLogoIconName}
\description{Sets the pixbuf to be displayed as logo in the about dialog.
If it is \code{NULL}, the default window icon set with
\code{\link{gtkWindowSetDefaultIcon}} will be used.}
\usage{gtkAboutDialogSetLogoIconName(object, icon.name = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}
\item{\verb{icon.name}}{an icon name, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
