\alias{gtkAboutDialogSetProgramName}
\name{gtkAboutDialogSetProgramName}
\title{gtkAboutDialogSetProgramName}
\description{Sets the name to display in the about dialog. 
If this is not set, it defaults to \code{gGetApplicationName()}.}
\usage{gtkAboutDialogSetProgramName(object, name)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAboutDialog}}]  a \code{\link{GtkAboutDialog}}}
\item{\code{name}}{[character]  the program name}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
