\alias{gtkAboutDialogSetWebsite}
\name{gtkAboutDialogSetWebsite}
\title{gtkAboutDialogSetWebsite}
\description{Sets the URL to use for the website link.}
\usage{gtkAboutDialogSetWebsite(object, website = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkAboutDialog}}]  a \verb{\link{GtkAboutDialog}}}
\item{\verb{website}}{[character]  a URL string starting with "http://"}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
