\alias{gtkAboutDialogSetWebsiteLabel}
\name{gtkAboutDialogSetWebsiteLabel}
\title{gtkAboutDialogSetWebsiteLabel}
\description{Sets the label to be used for the website link.
It defaults to the website URL.}
\usage{gtkAboutDialogSetWebsiteLabel(object, website.label = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAboutDialog}}]  a \code{\link{GtkAboutDialog}}}
\item{\code{website.label}}{[character]  the label used for the website link}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
