\alias{gtkAcceleratorGetDefaultModMask}
\name{gtkAcceleratorGetDefaultModMask}
\title{gtkAcceleratorGetDefaultModMask}
\description{Gets the value set by \code{\link{gtkAcceleratorSetDefaultModMask}}.}
\usage{gtkAcceleratorGetDefaultModMask()}
\value{[numeric]  the default accelerator modifier mask}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
