\alias{gtkAccelGroupActivate}
\name{gtkAccelGroupActivate}
\title{gtkAccelGroupActivate}
\description{\emph{undocumented
}}
\usage{gtkAccelGroupActivate(object, accel.quark, acceleratable, accel.key, accel.mods)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkAccelGroup}}] }
\item{\verb{accel.quark}}{[numeric] }
\item{\verb{acceleratable}}{[\verb{\link{GObject}}] }
\item{\verb{accel.key}}{[numeric] }
\item{\verb{accel.mods}}{[\code{\link{GdkModifierType}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
