\alias{gtkAccelGroupGetIsLocked}
\name{gtkAccelGroupGetIsLocked}
\title{gtkAccelGroupGetIsLocked}
\description{Locks are added and removed using \code{\link{gtkAccelGroupLock}} and
\code{\link{gtkAccelGroupUnlock}}.}
\usage{gtkAccelGroupGetIsLocked(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAccelGroup}}}}
\details{Since 2.14}
\value{[logical] \code{TRUE} if there are 1 or more locks on the \code{accel.group},
\code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
