\alias{gtkAccelGroupGetModifierMask}
\name{gtkAccelGroupGetModifierMask}
\title{gtkAccelGroupGetModifierMask}
\description{Gets a \code{\link{GdkModifierType}} representing the mask for this
\code{accel.group}. For example, \verb{GDK_CONTROL_MASK}, \verb{GDK_SHIFT_MASK}, etc.}
\usage{gtkAccelGroupGetModifierMask(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAccelGroup}}}}
\details{Since 2.14}
\value{[\code{\link{GdkModifierType}}] the modifier mask for this accel group.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
