\alias{gtkAccelGroupNew}
\name{gtkAccelGroupNew}
\title{gtkAccelGroupNew}
\description{Creates a new \code{\link{GtkAccelGroup}}.}
\usage{gtkAccelGroupNew()}
\value{[\code{\link{GtkAccelGroup}}]  a new \code{\link{GtkAccelGroup}} object}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
