\alias{gtkAccelGroupsFromObject}
\name{gtkAccelGroupsFromObject}
\title{gtkAccelGroupsFromObject}
\description{Gets a list of all accel groups which are attached to \code{object}.}
\usage{gtkAccelGroupsFromObject(object)}
\arguments{\item{\code{object}}{[\code{\link{GObject}}]         a \code{\link{GObject}}, usually a \code{\link{GtkWindow}} }}
\value{[list]  a list of all accel groups which are attached to \code{object}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
