\alias{gtkAccelLabelGetAccelWidth}
\name{gtkAccelLabelGetAccelWidth}
\title{gtkAccelLabelGetAccelWidth}
\description{Returns the width needed to display the accelerator key(s).
This is used by menus to align all of the \code{\link{GtkMenuItem}} widgets, and shouldn't
be needed by applications.}
\usage{gtkAccelLabelGetAccelWidth(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAccelLabel}}.}}
\value{[numeric] the width needed to display the accelerator key(s).}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
