\alias{gtkAccelLabelNew}
\name{gtkAccelLabelNew}
\title{gtkAccelLabelNew}
\description{Creates a new \verb{\link{GtkAccelLabel}}.}
\usage{gtkAccelLabelNew(string = NULL, show = TRUE)}
\arguments{\item{\verb{string}}{[character] the label string. Must be non-\code{NULL}.}}
\value{[\verb{\link{GtkWidget}}] a new \verb{\link{GtkAccelLabel}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
