\alias{gtkAccelLabelRefetch}
\name{gtkAccelLabelRefetch}
\title{gtkAccelLabelRefetch}
\description{Recreates the string representing the accelerator keys.
This should not be needed since the string is automatically updated whenever
accelerators are added or removed from the associated widget.}
\usage{gtkAccelLabelRefetch(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAccelLabel}}.}}
\value{[logical] always returns \code{FALSE}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
