\alias{gtkAccelLabelSetAccelWidget}
\name{gtkAccelLabelSetAccelWidget}
\title{gtkAccelLabelSetAccelWidget}
\description{Sets the widget to be monitored by this accelerator label.}
\usage{gtkAccelLabelSetAccelWidget(object, accel.widget)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkAccelLabel}}]  a \verb{\link{GtkAccelLabel}}}
\item{\verb{accel.widget}}{[\verb{\link{GtkWidget}}]  the widget to be monitored.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
