\alias{gtkAccelMapGet}
\name{gtkAccelMapGet}
\title{gtkAccelMapGet}
\description{Gets the singleton global \verb{\link{GtkAccelMap}} object. This object
is useful only for notification of changes to the accelerator
map via the ::changed signal; it isn't a parameter to the
other accelerator map functions.}
\usage{gtkAccelMapGet()}
\details{  Since  2.4}
\value{[\verb{\link{GtkAccelMap}}]  the global \verb{\link{GtkAccelMap}} object}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
