\alias{gtkAccelMapLoad}
\name{gtkAccelMapLoad}
\title{gtkAccelMapLoad}
\description{Parses a file previously saved with \code{\link{gtkAccelMapSave}} for
accelerator specifications, and propagates them accordingly.}
\usage{gtkAccelMapLoad(file.name)}
\arguments{\item{\verb{file.name}}{a file containing accelerator specifications,
  in the GLib file name encoding}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
