\alias{gtkAccelMapLoadFd}
\name{gtkAccelMapLoadFd}
\title{gtkAccelMapLoadFd}
\description{Filedescriptor variant of \code{\link{gtkAccelMapLoad}}.}
\usage{gtkAccelMapLoadFd(fd)}
\arguments{\item{\verb{fd}}{a valid readable file descriptor}}
\details{Note that the file descriptor will not be closed by this function.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
