\alias{gtkAccelMapLoadScanner}
\name{gtkAccelMapLoadScanner}
\title{gtkAccelMapLoadScanner}
\description{\code{GScanner} variant of \code{\link{gtkAccelMapLoad}}.}
\usage{gtkAccelMapLoadScanner(scanner)}
\arguments{\item{\code{scanner}}{[GScanner]  a \code{GScanner} which has already been provided with an input file}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
