\alias{gtkAccelMapUnlockPath}
\name{gtkAccelMapUnlockPath}
\title{gtkAccelMapUnlockPath}
\description{Undoes the last call to \code{\link{gtkAccelMapLockPath}} on this \code{accel.path}.
Refer to \code{\link{gtkAccelMapLockPath}} for information about accelerator path locking.}
\usage{gtkAccelMapUnlockPath(accel.path)}
\arguments{\item{\verb{accel.path}}{a valid accelerator path}}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
