\alias{gtkActionActivate}
\name{gtkActionActivate}
\title{gtkActionActivate}
\description{Emits the "activate" signal on the specified action, if it isn't 
insensitive. This gets called by the proxy widgets when they get 
activated.}
\usage{gtkActionActivate(object)}
\arguments{\item{\verb{object}}{the action object}}
\details{It can also be used to manually activate an action.
  Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
