\alias{gtkActionCreateIcon}
\name{gtkActionCreateIcon}
\title{gtkActionCreateIcon}
\description{This function is intended for use by action implementations to
create icons displayed in the proxy widgets.}
\usage{gtkActionCreateIcon(object, icon.size)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAction}}]  the action object}
\item{\code{icon.size}}{[\code{\link{GtkIconSize}}]  the size of the icon that should be created.}
}
\details{  Since  2.4}
\value{[\code{\link{GtkWidget}}]  a widget that displays the icon for this action.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
