\alias{gtkActionCreateMenuItem}
\name{gtkActionCreateMenuItem}
\title{gtkActionCreateMenuItem}
\description{Creates a menu item widget that proxies for the given action.}
\usage{gtkActionCreateMenuItem(object)}
\arguments{\item{\verb{object}}{the action object}}
\details{Since 2.4}
\value{[\code{\link{GtkWidget}}] a menu item connected to the action.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
