\alias{gtkActionDisconnectAccelerator}
\name{gtkActionDisconnectAccelerator}
\title{gtkActionDisconnectAccelerator}
\description{Undoes the effect of one call to \code{\link{gtkActionConnectAccelerator}}.}
\usage{gtkActionDisconnectAccelerator(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkAction}}]  a \code{\link{GtkAction}}}}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
