\alias{gtkActionDisconnectProxy}
\name{gtkActionDisconnectProxy}
\title{gtkActionDisconnectProxy}
\description{Disconnects a proxy widget from an action.  
Does \emph{not} destroy the widget, however.}
\usage{gtkActionDisconnectProxy(object, proxy)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAction}}]  the action object}
\item{\code{proxy}}{[\code{\link{GtkWidget}}]  the proxy widget}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
