\alias{gtkActionGetAccelClosure}
\name{gtkActionGetAccelClosure}
\title{gtkActionGetAccelClosure}
\description{Returns the accel closure for this action.}
\usage{gtkActionGetAccelClosure(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkAction}}]  the action object}}
\details{  Since  2.8}
\value{[\verb{\link{GClosure}}]  the accel closure for this action.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
