\alias{gtkActionGetAccelPath}
\name{gtkActionGetAccelPath}
\title{gtkActionGetAccelPath}
\description{Returns the accel path for this action.}
\usage{gtkActionGetAccelPath(object)}
\arguments{\item{\verb{object}}{the action object}}
\details{Since 2.6}
\value{[character] the accel path for this action, or \code{NULL}
if none is set.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
