\alias{gtkActionGetAlwaysShowImage}
\name{gtkActionGetAlwaysShowImage}
\title{gtkActionGetAlwaysShowImage}
\description{Returns whether \code{action}'s menu item proxies will ignore the
\verb{"gtk-menu-images"} setting and always show their image,
if available.}
\usage{gtkActionGetAlwaysShowImage(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAction}}}}
\details{Since 2.20}
\value{[logical] \code{TRUE} if the menu item proxies will always show their image}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
