\alias{gtkActionGetGicon}
\name{gtkActionGetGicon}
\title{gtkActionGetGicon}
\description{Gets the gicon of \code{action}.}
\usage{gtkActionGetGicon(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAction}}}}
\details{Since 2.16}
\value{[\code{\link{GIcon}}] The action's \code{\link{GIcon}} if one is set.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
