\alias{gtkActionGetIconName}
\name{gtkActionGetIconName}
\title{gtkActionGetIconName}
\description{Gets the icon name of \code{action}.}
\usage{gtkActionGetIconName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAction}}}}
\details{Since 2.16}
\value{[character] the icon name}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
