\alias{gtkActionGetIsImportant}
\name{gtkActionGetIsImportant}
\title{gtkActionGetIsImportant}
\description{Checks whether \code{action} is important or not}
\usage{gtkActionGetIsImportant(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAction}}}}
\details{Since 2.16}
\value{[logical] whether \code{action} is important}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
