\alias{gtkActionGetLabel}
\name{gtkActionGetLabel}
\title{gtkActionGetLabel}
\description{Gets the label text of \code{action}.}
\usage{gtkActionGetLabel(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAction}}}}
\details{Since 2.16}
\value{[character] the label text}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
