\alias{gtkActionGetName}
\name{gtkActionGetName}
\title{gtkActionGetName}
\description{Returns the name of the action.}
\usage{gtkActionGetName(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkAction}}]  the action object}}
\details{  Since  2.4}
\value{[character]  the name of the action.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
