\alias{gtkActionGetSensitive}
\name{gtkActionGetSensitive}
\title{gtkActionGetSensitive}
\description{Returns whether the action itself is sensitive. Note that this doesn't 
necessarily mean effective sensitivity. See \code{\link{gtkActionIsSensitive}} 
for that.}
\usage{gtkActionGetSensitive(object)}
\arguments{\item{\verb{object}}{the action object}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if the action itself is sensitive.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
