\alias{gtkActionGetShortLabel}
\name{gtkActionGetShortLabel}
\title{gtkActionGetShortLabel}
\description{Gets the short label text of \code{action}.}
\usage{gtkActionGetShortLabel(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAction}}}}
\details{Since 2.16}
\value{[character] the short label text.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
