\alias{gtkActionGetStockId}
\name{gtkActionGetStockId}
\title{gtkActionGetStockId}
\description{Gets the stock id of \code{action}.}
\usage{gtkActionGetStockId(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAction}}}}
\details{Since 2.16}
\value{[character] the stock id}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
