\alias{gtkActionGetTooltip}
\name{gtkActionGetTooltip}
\title{gtkActionGetTooltip}
\description{Gets the tooltip text of \code{action}.}
\usage{gtkActionGetTooltip(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAction}}}}
\details{Since 2.16}
\value{[character] the tooltip text}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
