\alias{gtkActionGetVisible}
\name{gtkActionGetVisible}
\title{gtkActionGetVisible}
\description{Returns whether the action itself is visible. Note that this doesn't 
necessarily mean effective visibility. See \code{\link{gtkActionIsSensitive}} 
for that.}
\usage{gtkActionGetVisible(object)}
\arguments{\item{\verb{object}}{the action object}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if the action itself is visible.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
