\alias{gtkActionGetVisibleVertical}
\name{gtkActionGetVisibleVertical}
\title{gtkActionGetVisibleVertical}
\description{Checks whether \code{action} is visible when horizontal}
\usage{gtkActionGetVisibleVertical(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAction}}}}
\details{Since 2.16}
\value{[logical] whether \code{action} is visible when horizontal}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
