\alias{gtkActionGroupGetAction}
\name{gtkActionGroupGetAction}
\title{gtkActionGroupGetAction}
\description{Looks up an action in the action group by name.}
\usage{gtkActionGroupGetAction(object, action.name)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkActionGroup}}]  the action group}
\item{\verb{action.name}}{[character]  the name of the action}
}
\details{  Since  2.4}
\value{[\verb{\link{GtkAction}}]  the action, or \code{NULL} if no action by that name exists}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
