\alias{gtkActionGroupGetName}
\name{gtkActionGroupGetName}
\title{gtkActionGroupGetName}
\description{Gets the name of the action group.}
\usage{gtkActionGroupGetName(object)}
\arguments{\item{\verb{object}}{the action group}}
\details{Since 2.4}
\value{[character] the name of the action group.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
