\alias{gtkActionGroupNew}
\name{gtkActionGroupNew}
\title{gtkActionGroupNew}
\description{Creates a new \verb{\link{GtkActionGroup}} object. The name of the action group
is used when associating keybindings 
with the actions.}
\usage{gtkActionGroupNew(name = NULL)}
\arguments{\item{\verb{name}}{[character]  the name of the action group.}}
\details{  Since  2.4}
\value{[\verb{\link{GtkActionGroup}}]  the new \verb{\link{GtkActionGroup}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
