\alias{gtkActionGroupSetSensitive}
\name{gtkActionGroupSetSensitive}
\title{gtkActionGroupSetSensitive}
\description{Changes the sensitivity of \code{action.group}}
\usage{gtkActionGroupSetSensitive(object, sensitive)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkActionGroup}}]  the action group}
\item{\verb{sensitive}}{[logical]  new sensitivity}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
