\alias{gtkActionGroupTranslateString}
\name{gtkActionGroupTranslateString}
\title{gtkActionGroupTranslateString}
\description{Translates a string using the specified \code{translate\Func()}. This
is mainly intended for language bindings.}
\usage{gtkActionGroupTranslateString(object, string)}
\arguments{
\item{\code{object}}{[\code{\link{GtkActionGroup}}]  a \code{\link{GtkActionGroup}}}
\item{\code{string}}{[character]  a string}
}
\details{  Since  2.6}
\value{[character]  the translation of \code{string}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
