\alias{gtkActionIsVisible}
\name{gtkActionIsVisible}
\title{gtkActionIsVisible}
\description{Returns whether the action is effectively visible.}
\usage{gtkActionIsVisible(object)}
\arguments{\item{\verb{object}}{the action object}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if the action and its associated action group
are both visible.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
