\alias{gtkActionSetAccelGroup}
\name{gtkActionSetAccelGroup}
\title{gtkActionSetAccelGroup}
\description{Sets the \verb{\link{GtkAccelGroup}} in which the accelerator for this action
will be installed.}
\usage{gtkActionSetAccelGroup(object, accel.group)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkAction}}]  the action object}
\item{\verb{accel.group}}{[\verb{\link{GtkAccelGroup}}]  a \verb{\link{GtkAccelGroup}} or \code{NULL}}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
