\alias{gtkActionSetGicon}
\name{gtkActionSetGicon}
\title{gtkActionSetGicon}
\description{Sets the icon of \code{action}.}
\usage{gtkActionSetGicon(object, icon)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAction}}}
\item{\verb{icon}}{the \code{\link{GIcon}} to set}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
