\alias{gtkActionSetIconName}
\name{gtkActionSetIconName}
\title{gtkActionSetIconName}
\description{Sets the icon name on \code{action}}
\usage{gtkActionSetIconName(object, icon.name)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAction}}}
\item{\verb{icon.name}}{the icon name to set}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
