\alias{gtkActionSetIsImportant}
\name{gtkActionSetIsImportant}
\title{gtkActionSetIsImportant}
\description{Sets whether the action is important, this attribute is used
primarily by toolbar items to decide whether to show a label
or not.}
\usage{gtkActionSetIsImportant(object, is.important)}
\arguments{
\item{\verb{object}}{the action object}
\item{\verb{is.important}}{\code{TRUE} to make the action important}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
