\alias{gtkActionSetLabel}
\name{gtkActionSetLabel}
\title{gtkActionSetLabel}
\description{Sets the label of \code{action}.}
\usage{gtkActionSetLabel(object, label)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAction}}}
\item{\verb{label}}{the label text to set}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
