\alias{gtkActionSetShortLabel}
\name{gtkActionSetShortLabel}
\title{gtkActionSetShortLabel}
\description{Sets a shorter label text on \code{action}.}
\usage{gtkActionSetShortLabel(object, short.label)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAction}}}
\item{\verb{short.label}}{the label text to set}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
