\alias{gtkActionSetStockId}
\name{gtkActionSetStockId}
\title{gtkActionSetStockId}
\description{Sets the stock id on \code{action}}
\usage{gtkActionSetStockId(object, stock.id)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAction}}}
\item{\verb{stock.id}}{the stock id}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
