\alias{gtkActionSetTooltip}
\name{gtkActionSetTooltip}
\title{gtkActionSetTooltip}
\description{Sets the tooltip text on \code{action}}
\usage{gtkActionSetTooltip(object, tooltip)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAction}}}
\item{\verb{tooltip}}{the tooltip text}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
