\alias{gtkActionSetVisibleHorizontal}
\name{gtkActionSetVisibleHorizontal}
\title{gtkActionSetVisibleHorizontal}
\description{Sets whether \code{action} is visible when horizontal}
\usage{gtkActionSetVisibleHorizontal(object, visible.horizontal)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAction}}}
\item{\verb{visible.horizontal}}{whether the action is visible horizontally}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
