\alias{gtkActionSetVisibleVertical}
\name{gtkActionSetVisibleVertical}
\title{gtkActionSetVisibleVertical}
\description{Sets whether \code{action} is visible when vertical}
\usage{gtkActionSetVisibleVertical(object, visible.vertical)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAction}}}
\item{\verb{visible.vertical}}{whether the action is visible vertically}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
