\alias{gtkActionUnblockActivate}
\name{gtkActionUnblockActivate}
\title{gtkActionUnblockActivate}
\description{Reenable activation signals from the action}
\usage{gtkActionUnblockActivate(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAction}}}}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
