\alias{gtkActivatableGetRelatedAction}
\name{gtkActivatableGetRelatedAction}
\title{gtkActivatableGetRelatedAction}
\description{Gets the related \code{\link{GtkAction}} for \code{activatable}.}
\usage{gtkActivatableGetRelatedAction(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkActivatable}}}}
\details{Since 2.16}
\value{[\code{\link{GtkAction}}] the related \code{\link{GtkAction}} if one is set.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
