\alias{gtkActivatableGetUseActionAppearance}
\name{gtkActivatableGetUseActionAppearance}
\title{gtkActivatableGetUseActionAppearance}
\description{Gets whether this activatable should reset its layout
and appearance when setting the related action or when
the action changes appearance.}
\usage{gtkActivatableGetUseActionAppearance(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkActivatable}}}}
\details{Since 2.16}
\value{[logical] whether \code{activatable} uses its actions appearance.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
