\alias{gtkAdjustmentGetLower}
\name{gtkAdjustmentGetLower}
\title{gtkAdjustmentGetLower}
\description{Retrieves the minimum value of the adjustment.}
\usage{gtkAdjustmentGetLower(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAdjustment}}}}
\details{Since 2.14}
\value{[numeric] The current minimum value of the adjustment.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
