\alias{gtkAdjustmentGetPageIncrement}
\name{gtkAdjustmentGetPageIncrement}
\title{gtkAdjustmentGetPageIncrement}
\description{Retrieves the page increment of the adjustment.}
\usage{gtkAdjustmentGetPageIncrement(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAdjustment}}}}
\details{Since 2.14}
\value{[numeric] The current page increment of the adjustment.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
