\alias{gtkAdjustmentGetPageSize}
\name{gtkAdjustmentGetPageSize}
\title{gtkAdjustmentGetPageSize}
\description{Retrieves the page size of the adjustment.}
\usage{gtkAdjustmentGetPageSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAdjustment}}}}
\details{Since 2.14}
\value{[numeric] The current page size of the adjustment.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
