\alias{gtkAdjustmentGetStepIncrement}
\name{gtkAdjustmentGetStepIncrement}
\title{gtkAdjustmentGetStepIncrement}
\description{Retrieves the step increment of the adjustment.}
\usage{gtkAdjustmentGetStepIncrement(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAdjustment}}}}
\details{Since 2.14}
\value{[numeric] The current step increment of the adjustment.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
