\alias{gtkAdjustmentGetUpper}
\name{gtkAdjustmentGetUpper}
\title{gtkAdjustmentGetUpper}
\description{Retrieves the maximum value of the adjustment.}
\usage{gtkAdjustmentGetUpper(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAdjustment}}}}
\details{Since 2.14}
\value{[numeric] The current maximum value of the adjustment.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
