\alias{gtkAdjustmentGetValue}
\name{gtkAdjustmentGetValue}
\title{gtkAdjustmentGetValue}
\description{Gets the current value of the adjustment. See
\code{\link{gtkAdjustmentSetValue}}.}
\usage{gtkAdjustmentGetValue(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAdjustment}}}}
\value{[numeric] The current value of the adjustment.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
