\alias{gtkAdjustmentNew}
\name{gtkAdjustmentNew}
\title{gtkAdjustmentNew}
\description{Creates a new \verb{\link{GtkAdjustment}}.}
\usage{gtkAdjustmentNew(value = NULL, lower = NULL, upper = NULL, step.incr = NULL, page.incr = NULL, page.size = NULL)}
\arguments{
\item{\verb{value}}{[numeric] the initial value.}
\item{\verb{lower}}{[numeric] the minimum value.}
\item{\verb{upper}}{[numeric] the maximum value.}
\item{\verb{step.incr}}{[numeric] the step increment.}
\item{\verb{page.incr}}{[numeric] the page increment.}
\item{\verb{page.size}}{[numeric] the page size.}
}
\value{[\verb{\link{GtkObject}}] a new \verb{\link{GtkAdjustment}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
