\alias{gtkAdjustmentSetPageIncrement}
\name{gtkAdjustmentSetPageIncrement}
\title{gtkAdjustmentSetPageIncrement}
\description{Sets the page increment of the adjustment.}
\usage{gtkAdjustmentSetPageIncrement(object, page.increment)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAdjustment}}}
\item{\verb{page.increment}}{the new page increment}
}
\details{See \code{\link{gtkAdjustmentSetLower}} about how to compress multiple
emissions of the "changed" signal when setting multiple adjustment
properties.
  Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
