\alias{gtkAdjustmentSetPageSize}
\name{gtkAdjustmentSetPageSize}
\title{gtkAdjustmentSetPageSize}
\description{Sets the page size of the adjustment.}
\usage{gtkAdjustmentSetPageSize(object, page.size)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAdjustment}}}
\item{\verb{page.size}}{the new page size}
}
\details{See \code{\link{gtkAdjustmentSetLower}} about how to compress multiple
emissions of the "changed" signal when setting multiple adjustment
properties.
  Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
