\alias{gtkAdjustmentSetStepIncrement}
\name{gtkAdjustmentSetStepIncrement}
\title{gtkAdjustmentSetStepIncrement}
\description{Sets the step increment of the adjustment.}
\usage{gtkAdjustmentSetStepIncrement(object, step.increment)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAdjustment}}}
\item{\verb{step.increment}}{the new step increment}
}
\details{See \code{\link{gtkAdjustmentSetLower}} about how to compress multiple
emissions of the "changed" signal when setting multiple adjustment
properties.
  Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
