\alias{gtkAdjustmentValueChanged}
\name{gtkAdjustmentValueChanged}
\title{gtkAdjustmentValueChanged}
\description{Emits a "value_changed" signal from the \verb{\link{GtkAdjustment}}.
This is typically called by the owner of the \verb{\link{GtkAdjustment}} after it has
changed the \verb{\link{GtkAdjustment}} value field.}
\usage{gtkAdjustmentValueChanged(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkAdjustment}}] a \verb{\link{GtkAdjustment}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
