\alias{gtkAlignmentSetPadding}
\name{gtkAlignmentSetPadding}
\title{gtkAlignmentSetPadding}
\description{Sets the padding on the different sides of the widget.
The padding adds blank space to the sides of the widget. For instance,
this can be used to indent the child widget towards the right by adding
padding on the left.}
\usage{gtkAlignmentSetPadding(object, padding.top, padding.bottom, padding.left, padding.right)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkAlignment}}]  a \verb{\link{GtkAlignment}}}
\item{\verb{padding.top}}{[numeric]  the padding at the top of the widget}
\item{\verb{padding.bottom}}{[numeric]  the padding at the bottom of the widget}
\item{\verb{padding.left}}{[numeric]  the padding at the left of the widget}
\item{\verb{padding.right}}{[numeric]  the padding at the right of the widget.}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
