\name{GtkAllocation}
\alias{GtkAllocation}
\title{GtkAllocation}
\description{
A \code{GtkAllocation} is just a fancy name for a \code{\link{GdkRectangle}} 
that represents the region allocated to a widget by its parent.
}
\seealso{
\code{\link{GdkRectangle}}
}
\author{Michael Lawrence}
\keyword{interface}
\keyword{internal}
