\alias{gtkAssistantAddActionWidget}
\name{gtkAssistantAddActionWidget}
\title{gtkAssistantAddActionWidget}
\description{Adds a widget to the action area of a \verb{\link{GtkAssistant}}.}
\usage{gtkAssistantAddActionWidget(object, child)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkAssistant}}]  a \verb{\link{GtkAssistant}}}
\item{\verb{child}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
