\alias{gtkAssistantGetCurrentPage}
\name{gtkAssistantGetCurrentPage}
\title{gtkAssistantGetCurrentPage}
\description{Returns the page number of the current page}
\usage{gtkAssistantGetCurrentPage(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAssistant}}}}
\details{Since 2.10}
\value{[integer] The index (starting from 0) of the current page in
the \code{assistant}, if the \code{assistant} has no pages, -1 will be returned}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
