\alias{gtkAssistantGetNPages}
\name{gtkAssistantGetNPages}
\title{gtkAssistantGetNPages}
\description{Returns the number of pages in the \code{assistant}}
\usage{gtkAssistantGetNPages(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAssistant}}}}
\details{Since 2.10}
\value{[integer] The number of pages in the \code{assistant}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
