\alias{gtkAssistantGetPageComplete}
\name{gtkAssistantGetPageComplete}
\title{gtkAssistantGetPageComplete}
\description{Gets whether \code{page} is complete..}
\usage{gtkAssistantGetPageComplete(object, page)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAssistant}}]  a \code{\link{GtkAssistant}}}
\item{\code{page}}{[\code{\link{GtkWidget}}]  a page of \code{assistant}}
}
\details{  Since  2.10}
\value{[logical] \code{TRUE} if \code{page} is complete.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
