\alias{gtkAssistantGetPageSideImage}
\name{gtkAssistantGetPageSideImage}
\title{gtkAssistantGetPageSideImage}
\description{Gets the header image for \code{page}.}
\usage{gtkAssistantGetPageSideImage(object, page)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAssistant}}]  a \code{\link{GtkAssistant}}}
\item{\code{page}}{[\code{\link{GtkWidget}}]  a page of \code{assistant}}
}
\details{  Since  2.10}
\value{[\code{\link{GdkPixbuf}}]  the side image for \code{page}, or \code{NULL}
if there's no side image for the page.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
