\alias{gtkAssistantNew}
\name{gtkAssistantNew}
\title{gtkAssistantNew}
\description{Creates a new \verb{\link{GtkAssistant}}.}
\usage{gtkAssistantNew(show = TRUE)}
\details{  Since  2.10}
\value{[\verb{\link{GtkWidget}}]  a newly created \verb{\link{GtkAssistant}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
