\alias{gtkAssistantPrependPage}
\name{gtkAssistantPrependPage}
\title{gtkAssistantPrependPage}
\description{Prepends a page to the \code{assistant}.}
\usage{gtkAssistantPrependPage(object, page)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkAssistant}}]  a \verb{\link{GtkAssistant}}}
\item{\verb{page}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}}
}
\details{  Since  2.10}
\value{[integer]  the index (starting at 0) of the inserted page}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
