\alias{gtkAssistantRemoveActionWidget}
\name{gtkAssistantRemoveActionWidget}
\title{gtkAssistantRemoveActionWidget}
\description{Removes a widget from the action area of a \verb{\link{GtkAssistant}}.}
\usage{gtkAssistantRemoveActionWidget(object, child)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkAssistant}}]  a \verb{\link{GtkAssistant}}}
\item{\verb{child}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
