\alias{gtkAssistantSetPageHeaderImage}
\name{gtkAssistantSetPageHeaderImage}
\title{gtkAssistantSetPageHeaderImage}
\description{Sets a header image for \code{page}. This image is displayed in the header
area of the assistant when \code{page} is the current page.}
\usage{gtkAssistantSetPageHeaderImage(object, page, pixbuf = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkAssistant}}]  a \verb{\link{GtkAssistant}}}
\item{\verb{page}}{[\verb{\link{GtkWidget}}]  a page of \code{assistant}}
\item{\verb{pixbuf}}{[\verb{\link{GdkPixbuf}}]  the new header image \code{page}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
