\alias{gtkAssistantSetPageSideImage}
\name{gtkAssistantSetPageSideImage}
\title{gtkAssistantSetPageSideImage}
\description{Sets a header image for \code{page}. This image is displayed in the side
area of the assistant when \code{page} is the current page.}
\usage{gtkAssistantSetPageSideImage(object, page, pixbuf = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkAssistant}}]  a \verb{\link{GtkAssistant}}}
\item{\verb{page}}{[\verb{\link{GtkWidget}}]  a page of \code{assistant}}
\item{\verb{pixbuf}}{[\verb{\link{GdkPixbuf}}]  the new header image \code{page}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
