\alias{gtkAssistantSetPageTitle}
\name{gtkAssistantSetPageTitle}
\title{gtkAssistantSetPageTitle}
\description{Sets a title for \code{page}. The title is displayed in the header
area of the assistant when \code{page} is the current page.}
\usage{gtkAssistantSetPageTitle(object, page, title)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAssistant}}]  a \code{\link{GtkAssistant}}}
\item{\code{page}}{[\code{\link{GtkWidget}}]  a page of \code{assistant}}
\item{\code{title}}{[character]  the new title for \code{page}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
