\alias{gtkBorderNew}
\name{gtkBorderNew}
\title{gtkBorderNew}
\description{Allocates a new \code{\link{GtkBorder}} structure and initializes its elements to zero.}
\usage{gtkBorderNew()}
\details{Since 2.14}
\value{[\code{\link{GtkBorder}}] a new empty \code{\link{GtkBorder}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
