\alias{gtkBoxGetHomogeneous}
\name{gtkBoxGetHomogeneous}
\title{gtkBoxGetHomogeneous}
\description{Returns whether the box is homogeneous (all children are the
same size). See \code{\link{gtkBoxSetHomogeneous}}.}
\usage{gtkBoxGetHomogeneous(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkBox}}}}
\value{[logical] \code{TRUE} if the box is homogeneous.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
