\alias{gtkBoxSetSpacing}
\name{gtkBoxSetSpacing}
\title{gtkBoxSetSpacing}
\description{Sets the \code{"spacing"} property of \code{box}, which is the 
number of pixels to place between children of \code{box}.}
\usage{gtkBoxSetSpacing(object, spacing)}
\arguments{
\item{\code{object}}{[\code{\link{GtkBox}}]  a \code{\link{GtkBox}}}
\item{\code{spacing}}{[integer]  the number of pixels to put between children}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
