\alias{gtkBuildableAddChild}
\name{gtkBuildableAddChild}
\title{gtkBuildableAddChild}
\description{Adds a child to \code{buildable}. \code{type} is an optional string
describing how the child should be added.}
\usage{gtkBuildableAddChild(object, builder, child, type)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkBuildable}}]  a \verb{\link{GtkBuildable}}}
\item{\verb{builder}}{[\verb{\link{GtkBuilder}}]  a \verb{\link{GtkBuilder}}}
\item{\verb{child}}{[\verb{\link{GObject}}]  child to add}
\item{\verb{type}}{[character]  kind of child or \code{NULL}}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
