\alias{gtkBuildableGetInternalChild}
\name{gtkBuildableGetInternalChild}
\title{gtkBuildableGetInternalChild}
\description{Get the internal child called \code{childname} of the \code{buildable} object.}
\usage{gtkBuildableGetInternalChild(object, builder, childname)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkBuildable}}]  a \verb{\link{GtkBuildable}}}
\item{\verb{builder}}{[\verb{\link{GtkBuilder}}]  a \verb{\link{GtkBuilder}}}
\item{\verb{childname}}{[character]  name of child}
}
\details{  Since  2.12}
\value{[\verb{\link{GObject}}]  the internal child of the buildable object }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
