\alias{gtkBuildableSetBuildableProperty}
\name{gtkBuildableSetBuildableProperty}
\title{gtkBuildableSetBuildableProperty}
\description{Sets the property name \code{name} to \code{value} on the \code{buildable} object.}
\usage{gtkBuildableSetBuildableProperty(object, builder, name, value)}
\arguments{
\item{\code{object}}{[\code{\link{GtkBuildable}}]  a \code{\link{GtkBuildable}}}
\item{\code{builder}}{[\code{\link{GtkBuilder}}]  a \code{\link{GtkBuilder}}}
\item{\code{name}}{[character]  name of property}
\item{\code{value}}{[R object]  value of property}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
