\alias{gtkBuildableSetName}
\name{gtkBuildableSetName}
\title{gtkBuildableSetName}
\description{Sets the name of the \code{buildable} object.}
\usage{gtkBuildableSetName(object, name)}
\arguments{
\item{\code{object}}{[\code{\link{GtkBuildable}}]  a \code{\link{GtkBuildable}}}
\item{\code{name}}{[character]  name to set}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
