\alias{gtkBuilderGetTranslationDomain}
\name{gtkBuilderGetTranslationDomain}
\title{gtkBuilderGetTranslationDomain}
\description{Gets the translation domain of \code{builder}.}
\usage{gtkBuilderGetTranslationDomain(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkBuilder}}]  a \code{\link{GtkBuilder}}}}
\details{  Since  2.12}
\value{[character]  the translation domain.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
