\alias{gtkBuilderNew}
\name{gtkBuilderNew}
\title{gtkBuilderNew}
\description{Creates a new builder object.}
\usage{gtkBuilderNew()}
\details{  Since  2.12}
\value{[\code{\link{GtkBuilder}}]  a new \code{\link{GtkBuilder}} object}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
