\alias{gtkBuilderSetTranslationDomain}
\name{gtkBuilderSetTranslationDomain}
\title{gtkBuilderSetTranslationDomain}
\description{Sets the translation domain of \code{builder}. 
See \code{:translation-domain}.}
\usage{gtkBuilderSetTranslationDomain(object, domain)}
\arguments{
\item{\code{object}}{[\code{\link{GtkBuilder}}]  a \code{\link{GtkBuilder}}}
\item{\code{domain}}{[character]  the translation domain or \code{NULL}}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
