\alias{gtkButtonBoxGetChildSecondary}
\name{gtkButtonBoxGetChildSecondary}
\title{gtkButtonBoxGetChildSecondary}
\description{Returns whether \code{child} should appear in a secondary group of children.}
\usage{gtkButtonBoxGetChildSecondary(object, child)}
\arguments{
\item{\code{object}}{[\code{\link{GtkButtonBox}}]  a \code{\link{GtkButtonBox}}}
\item{\code{child}}{[\code{\link{GtkWidget}}]  a child of \code{widget} }
}
\details{  Since  2.4}
\value{[logical]  whether \code{child} should appear in a secondary group of children.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
