\alias{gtkButtonBoxGetLayout}
\name{gtkButtonBoxGetLayout}
\title{gtkButtonBoxGetLayout}
\description{Retrieves the method being used to arrange the buttons in a button box.}
\usage{gtkButtonBoxGetLayout(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkButtonBox}}] a \verb{\link{GtkButtonBox}}.}}
\value{[\code{\link{GtkButtonBoxStyle}}] the method used to layout buttons in \code{widget}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
