\alias{gtkButtonBoxSetLayout}
\name{gtkButtonBoxSetLayout}
\title{gtkButtonBoxSetLayout}
\description{Changes the way buttons are arranged in their container.}
\usage{gtkButtonBoxSetLayout(object, layout.style)}
\arguments{
\item{\code{object}}{[\code{\link{GtkButtonBox}}] a \code{\link{GtkButtonBox}}.}
\item{\code{layout.style}}{[\code{\link{GtkButtonBoxStyle}}] the new layout style.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
