\alias{gtkButtonEnter}
\name{gtkButtonEnter}
\title{gtkButtonEnter}
\description{Emits a \code{\link{gtkButtonEnter}} signal to the given \verb{\link{GtkButton}}.}
\usage{gtkButtonEnter(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkButton}}] The \verb{\link{GtkButton}} you want to send the signal to.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
