\alias{gtkButtonGetFocusOnClick}
\name{gtkButtonGetFocusOnClick}
\title{gtkButtonGetFocusOnClick}
\description{Returns whether the button grabs focus when it is clicked with the mouse.
See \code{\link{gtkButtonSetFocusOnClick}}.}
\usage{gtkButtonGetFocusOnClick(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkButton}}}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if the button grabs focus when it is clicked with
the mouse.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
