\alias{gtkButtonGetImage}
\name{gtkButtonGetImage}
\title{gtkButtonGetImage}
\description{Gets the widget that is currenty set as the image of \code{button}.
This may have been explicitly set by \code{\link{gtkButtonSetImage}}
or constructed by \code{\link{gtkButtonNewFromStock}}.}
\usage{gtkButtonGetImage(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkButton}}}}
\details{Since 2.6}
\value{[\code{\link{GtkWidget}}] a \code{\link{GtkWidget}} or \code{NULL} in case there is no image}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
