\alias{gtkButtonGetImagePosition}
\name{gtkButtonGetImagePosition}
\title{gtkButtonGetImagePosition}
\description{Gets the position of the image relative to the text 
inside the button.}
\usage{gtkButtonGetImagePosition(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkButton}}}}
\details{Since 2.10}
\value{[\code{\link{GtkPositionType}}] the position}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
