\alias{gtkButtonGetUseStock}
\name{gtkButtonGetUseStock}
\title{gtkButtonGetUseStock}
\description{Returns whether the button label is a stock item.}
\usage{gtkButtonGetUseStock(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkButton}}}}
\value{[logical] \code{TRUE} if the button label is used to
select a stock item instead of being
used directly as the label text.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
