\alias{gtkButtonGetUseUnderline}
\name{gtkButtonGetUseUnderline}
\title{gtkButtonGetUseUnderline}
\description{Returns whether an embedded underline in the button label indicates a
mnemonic. See \code{\link{gtkButtonSetUseUnderline}}.}
\usage{gtkButtonGetUseUnderline(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkButton}}}}
\value{[logical] \code{TRUE} if an embedded underline in the button label
indicates the mnemonic accelerator keys.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
