\alias{gtkButtonLeave}
\name{gtkButtonLeave}
\title{gtkButtonLeave}
\description{Emits a \code{\link{gtkButtonLeave}} signal to the given \verb{\link{GtkButton}}.}
\usage{gtkButtonLeave(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkButton}}] The \verb{\link{GtkButton}} you want to send the signal to.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
