\alias{gtkButtonNewWithLabel}
\name{gtkButtonNewWithLabel}
\title{gtkButtonNewWithLabel}
\description{Creates a \verb{\link{GtkButton}} widget with a \verb{\link{GtkLabel}} child containing the given
text.}
\usage{gtkButtonNewWithLabel(label, show = TRUE)}
\arguments{\item{\verb{label}}{[character] The text you want the \verb{\link{GtkLabel}} to hold.}}
\value{[\verb{\link{GtkWidget}}] The newly created \verb{\link{GtkButton}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
