\alias{gtkButtonReleased}
\name{gtkButtonReleased}
\title{gtkButtonReleased}
\description{Emits a \code{\link{gtkButtonReleased}} signal to the given \verb{\link{GtkButton}}.}
\usage{gtkButtonReleased(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkButton}}] The \verb{\link{GtkButton}} you want to send the signal to.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
