\alias{gtkButtonSetAlignment}
\name{gtkButtonSetAlignment}
\title{gtkButtonSetAlignment}
\description{Sets the alignment of the child. This property has no effect unless 
the child is a \code{\link{GtkMisc}} or a \code{GtkAligment}.}
\usage{gtkButtonSetAlignment(object, xalign, yalign)}
\arguments{
\item{\code{object}}{[\code{\link{GtkButton}}]  a \code{\link{GtkButton}}}
\item{\code{xalign}}{[numeric]  the horizontal position of the child, 0.0 is left aligned, 
  1.0 is right aligned}
\item{\code{yalign}}{[numeric]  the vertical position of the child, 0.0 is top aligned, 
  1.0 is bottom aligned}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
