\alias{gtkButtonSetImagePosition}
\name{gtkButtonSetImagePosition}
\title{gtkButtonSetImagePosition}
\description{Sets the position of the image relative to the text 
inside the button.}
\usage{gtkButtonSetImagePosition(object, position)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkButton}}]  a \verb{\link{GtkButton}}}
\item{\verb{position}}{[\code{\link{GtkPositionType}}]  the position}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
