\alias{gtkButtonSetLabel}
\name{gtkButtonSetLabel}
\title{gtkButtonSetLabel}
\description{Sets the text of the label of the button to \code{str}. This text is
also used to select the stock item if \code{\link{gtkButtonSetUseStock}}
is used.}
\usage{gtkButtonSetLabel(object, label)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkButton}}}
\item{\verb{label}}{a string}
}
\details{This will also clear any previously set labels.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
