\alias{gtkButtonSetRelief}
\name{gtkButtonSetRelief}
\title{gtkButtonSetRelief}
\description{Sets the relief style of the edges of the given \code{\link{GtkButton}} widget.
Three styles exist, GTK\_RELIEF\_NORMAL, GTK\_RELIEF\_HALF, GTK\_RELIEF\_NONE.
The default style is, as one can guess, GTK\_RELIEF\_NORMAL.}
\usage{gtkButtonSetRelief(object, newstyle)}
\arguments{
\item{\code{object}}{[\code{\link{GtkButton}}] The \code{\link{GtkButton}} you want to set relief styles of.}
\item{\code{newstyle}}{[\code{\link{GtkReliefStyle}}] The GtkReliefStyle as described above.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
