\alias{gtkButtonSetUseUnderline}
\name{gtkButtonSetUseUnderline}
\title{gtkButtonSetUseUnderline}
\description{If true, an underline in the text of the button label indicates
the next character should be used for the mnemonic accelerator key.}
\usage{gtkButtonSetUseUnderline(object, use.underline)}
\arguments{
\item{\code{object}}{[\code{\link{GtkButton}}]  a \code{\link{GtkButton}}}
\item{\code{use.underline}}{[logical]  \code{TRUE} if underlines in the text indicate mnemonics}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
