\alias{gtkCalendarClearMarks}
\name{gtkCalendarClearMarks}
\title{gtkCalendarClearMarks}
\description{Remove all visual markers.}
\usage{gtkCalendarClearMarks(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkCalendar}}]  a \code{\link{GtkCalendar}}}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
