\alias{gtkCalendarDisplayOptions}
\name{gtkCalendarDisplayOptions}
\title{gtkCalendarDisplayOptions}
\description{
Sets display options (whether to display the heading and the month headings).
\strong{WARNING: \code{gtk_calendar_display_options} is deprecated and should not be used in newly-written code. }
}
\usage{gtkCalendarDisplayOptions(object, flags)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCalendar}}]  a \code{\link{GtkCalendar}}.}
\item{\code{flags}}{[\code{\link{GtkCalendarDisplayOptions}}]  the display options to set.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
