\alias{gtkCalendarFreeze}
\name{gtkCalendarFreeze}
\title{gtkCalendarFreeze}
\description{
Does nothing. Previously locked the display of the calendar until
it was thawed with \code{\link{gtkCalendarThaw}}.
\strong{WARNING: \code{gtk_calendar_freeze} is deprecated and should not be used in newly-written code.  
Does nothing. Previously locked the display of the calendar until
it was thawed with \code{\link{gtkCalendarThaw}}.  }
}
\usage{gtkCalendarFreeze(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkCalendar}}]  a \code{\link{GtkCalendar}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
