\alias{gtkCalendarGetDetailHeightRows}
\name{gtkCalendarGetDetailHeightRows}
\title{gtkCalendarGetDetailHeightRows}
\description{Queries the height of detail cells, in rows.
See \verb{"detail-width-chars"}.}
\usage{gtkCalendarGetDetailHeightRows(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkCalendar}}.}}
\details{Since 2.14}
\value{[integer] The height of detail cells, in rows.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
