\alias{gtkCalendarGetDetailWidthChars}
\name{gtkCalendarGetDetailWidthChars}
\title{gtkCalendarGetDetailWidthChars}
\description{Queries the width of detail cells, in characters.
See \verb{"detail-width-chars"}.}
\usage{gtkCalendarGetDetailWidthChars(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkCalendar}}.}}
\details{Since 2.14}
\value{[integer] The width of detail cells, in characters.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
