\alias{gtkCalendarGetDisplayOptions}
\name{gtkCalendarGetDisplayOptions}
\title{gtkCalendarGetDisplayOptions}
\description{Returns the current display options of \code{calendar}.}
\usage{gtkCalendarGetDisplayOptions(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkCalendar}}}}
\details{Since 2.4}
\value{[\code{\link{GtkCalendarDisplayOptions}}] the display options.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
