\alias{gtkCalendarNew}
\name{gtkCalendarNew}
\title{gtkCalendarNew}
\description{Creates a new calendar, with the current date being selected.}
\usage{gtkCalendarNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}]  a newly \verb{\link{GtkCalendar}} widget}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
