\alias{gtkCalendarSelectMonth}
\name{gtkCalendarSelectMonth}
\title{gtkCalendarSelectMonth}
\description{Shifts the calendar to a different month.}
\usage{gtkCalendarSelectMonth(object, month, year)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCalendar}}]  a \verb{\link{GtkCalendar}}}
\item{\verb{month}}{[numeric]  a month number between 0 and 11.}
\item{\verb{year}}{[numeric]  the year the month is in.}
}
\value{[logical] \code{TRUE}, always}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
