\alias{gtkCalendarSetDetailHeightRows}
\name{gtkCalendarSetDetailHeightRows}
\title{gtkCalendarSetDetailHeightRows}
\description{Updates the height of detail cells.
See \verb{"detail-height-rows"}.}
\usage{gtkCalendarSetDetailHeightRows(object, rows)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkCalendar}}.}
\item{\verb{rows}}{detail height in rows.}
}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
