\alias{gtkCalendarSetDetailWidthChars}
\name{gtkCalendarSetDetailWidthChars}
\title{gtkCalendarSetDetailWidthChars}
\description{Updates the width of detail cells.
See \verb{"detail-width-chars"}.}
\usage{gtkCalendarSetDetailWidthChars(object, chars)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkCalendar}}.}
\item{\verb{chars}}{detail width in characters.}
}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
