\alias{gtkCalendarThaw}
\name{gtkCalendarThaw}
\title{gtkCalendarThaw}
\description{
Does nothing. Previously defrosted a calendar; all the changes made
since the last \code{\link{gtkCalendarFreeze}} were displayed.
\strong{WARNING: \code{gtk_calendar_thaw} is deprecated and should not be used in newly-written code.  
Does nothing. Previously defrosted a calendar; all the changes made
since the last \code{\link{gtkCalendarFreeze}} were displayed.  }
}
\usage{gtkCalendarThaw(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkCalendar}}]  a \code{\link{GtkCalendar}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
