\alias{gtkCellEditableEditingDone}
\name{gtkCellEditableEditingDone}
\title{gtkCellEditableEditingDone}
\description{Emits the "editing_done" signal.  This signal is a sign for the cell renderer
to update its value from the cell.}
\usage{gtkCellEditableEditingDone(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkCellEditable}}]  A \code{GtkTreeEditable}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
