\alias{gtkCellEditableRemoveWidget}
\name{gtkCellEditableRemoveWidget}
\title{gtkCellEditableRemoveWidget}
\description{Emits the \code{\link{gtkCellEditableRemoveWidget}} signal.}
\usage{gtkCellEditableRemoveWidget(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkCellEditable}}]  A \code{GtkTreeEditable}}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
