\alias{gtkCellEditableStartEditing}
\name{gtkCellEditableStartEditing}
\title{gtkCellEditableStartEditing}
\description{Begins editing on a \code{cell.editable}. \code{event} is the \verb{\link{GdkEvent}} that began 
the editing process. It may be \code{NULL}, in the instance that editing was 
initiated through programatic means.}
\usage{gtkCellEditableStartEditing(object, event = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCellEditable}}]  A \verb{\link{GtkCellEditable}}}
\item{\verb{event}}{[\verb{\link{GdkEvent}}]  A \verb{\link{GdkEvent}}, or \code{NULL}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
