\alias{gtkCellLayoutClear}
\name{gtkCellLayoutClear}
\title{gtkCellLayoutClear}
\description{Unsets all the mappings on all renderers on \code{cell.layout} and
removes all renderers from \code{cell.layout}.}
\usage{gtkCellLayoutClear(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkCellLayout}}.}}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
