\alias{gtkCellLayoutClearAttributes}
\name{gtkCellLayoutClearAttributes}
\title{gtkCellLayoutClearAttributes}
\description{Clears all existing attributes previously set with
\code{\link{gtkCellLayoutSetAttributes}}.}
\usage{gtkCellLayoutClearAttributes(object, cell)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCellLayout}}]  A \code{\link{GtkCellLayout}}.}
\item{\code{cell}}{[\code{\link{GtkCellRenderer}}]  A \code{\link{GtkCellRenderer}} to clear the attribute mapping on.}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
