\alias{gtkCellRendererAccelNew}
\name{gtkCellRendererAccelNew}
\title{gtkCellRendererAccelNew}
\description{Creates a new \code{\link{GtkCellRendererAccel}}.}
\usage{gtkCellRendererAccelNew()}
\details{  Since  2.10}
\value{[\code{\link{GtkCellRenderer}}]  the new cell renderer}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
