\alias{gtkCellRendererComboNew}
\name{gtkCellRendererComboNew}
\title{gtkCellRendererComboNew}
\description{Creates a new \verb{\link{GtkCellRendererCombo}}. 
Adjust how text is drawn using object properties. 
Object properties can be set globally (with \code{\link{gObjectSet}}). 
Also, with \verb{\link{GtkTreeViewColumn}}, you can bind a property to a value 
in a \verb{\link{GtkTreeModel}}. For example, you can bind the "text" property 
on the cell renderer to a string value in the model, thus rendering 
a different string in each row of the \verb{\link{GtkTreeView}}.}
\usage{gtkCellRendererComboNew()}
\details{  Since  2.6}
\value{[\verb{\link{GtkCellRenderer}}]  the new cell renderer}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
