\alias{gtkCellRendererGetAlignment}
\name{gtkCellRendererGetAlignment}
\title{gtkCellRendererGetAlignment}
\description{Fills in \code{xalign} and \code{yalign} with the appropriate values of \code{cell}.}
\usage{gtkCellRendererGetAlignment(object, xalign, yalign)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkCellRenderer}}}
\item{\verb{xalign}}{location to fill in with the x alignment of the cell, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
\item{\verb{yalign}}{location to fill in with the y alignment of the cell, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
