\alias{gtkCellRendererGetFixedSize}
\name{gtkCellRendererGetFixedSize}
\title{gtkCellRendererGetFixedSize}
\description{Fills in \code{width} and \code{height} with the appropriate size of \code{cell}.}
\usage{gtkCellRendererGetFixedSize(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkCellRenderer}}]  A \verb{\link{GtkCellRenderer}}}}
\value{
A list containing the following elements:
\item{\verb{width}}{[integer]  location to fill in with the fixed width of the widget, or \code{NULL}}
\item{\verb{height}}{[integer]  location to fill in with the fixed height of the widget, or \code{NULL}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
