\alias{gtkCellRendererGetPadding}
\name{gtkCellRendererGetPadding}
\title{gtkCellRendererGetPadding}
\description{Fills in \code{xpad} and \code{ypad} with the appropriate values of \code{cell}.}
\usage{gtkCellRendererGetPadding(object, xpad, ypad)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkCellRenderer}}}
\item{\verb{xpad}}{location to fill in with the x padding of the cell, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
\item{\verb{ypad}}{location to fill in with the y padding of the cell, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
