\alias{gtkCellRendererGetSensitive}
\name{gtkCellRendererGetSensitive}
\title{gtkCellRendererGetSensitive}
\description{Returns the cell renderer's sensitivity.}
\usage{gtkCellRendererGetSensitive(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkCellRenderer}}}}
\details{Since 2.18}
\value{[logical] \code{TRUE} if the cell renderer is sensitive}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
