\alias{gtkCellRendererGetVisible}
\name{gtkCellRendererGetVisible}
\title{gtkCellRendererGetVisible}
\description{Returns the cell renderer's visibility.}
\usage{gtkCellRendererGetVisible(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkCellRenderer}}}}
\details{Since 2.18}
\value{[logical] \code{TRUE} if the cell renderer is visible}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
