\alias{gtkCellRendererPixbufNew}
\name{gtkCellRendererPixbufNew}
\title{gtkCellRendererPixbufNew}
\description{Creates a new \verb{\link{GtkCellRendererPixbuf}}. Adjust rendering
parameters using object properties. Object properties can be set
globally (with \code{\link{gObjectSet}}). Also, with \verb{\link{GtkTreeViewColumn}}, you
can bind a property to a value in a \verb{\link{GtkTreeModel}}. For example, you
can bind the "pixbuf" property on the cell renderer to a pixbuf value
in the model, thus rendering a different image in each row of the
\verb{\link{GtkTreeView}}.}
\usage{gtkCellRendererPixbufNew()}
\value{[\verb{\link{GtkCellRenderer}}]  the new cell renderer}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
