\alias{gtkCellRendererSetAlignment}
\name{gtkCellRendererSetAlignment}
\title{gtkCellRendererSetAlignment}
\description{Sets the renderer's alignment within its available space.}
\usage{gtkCellRendererSetAlignment(object, xalign, yalign)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkCellRenderer}}}
\item{\verb{xalign}}{the x alignment of the cell renderer}
\item{\verb{yalign}}{the y alignment of the cell renderer}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
