\alias{gtkCellRendererSetPadding}
\name{gtkCellRendererSetPadding}
\title{gtkCellRendererSetPadding}
\description{Sets the renderer's padding.}
\usage{gtkCellRendererSetPadding(object, xpad, ypad)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkCellRenderer}}}
\item{\verb{xpad}}{the x padding of the cell renderer}
\item{\verb{ypad}}{the y padding of the cell renderer}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
