\alias{gtkCellRendererSetSensitive}
\name{gtkCellRendererSetSensitive}
\title{gtkCellRendererSetSensitive}
\description{Sets the cell renderer's sensitivity.}
\usage{gtkCellRendererSetSensitive(object, sensitive)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkCellRenderer}}}
\item{\verb{sensitive}}{the sensitivity of the cell}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
