\alias{gtkCellRendererSetVisible}
\name{gtkCellRendererSetVisible}
\title{gtkCellRendererSetVisible}
\description{Sets the cell renderer's visibility.}
\usage{gtkCellRendererSetVisible(object, visible)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkCellRenderer}}}
\item{\verb{visible}}{the visibility of the cell}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
