\alias{gtkCellRendererSpinnerNew}
\name{gtkCellRendererSpinnerNew}
\title{gtkCellRendererSpinnerNew}
\description{Returns a new cell renderer which will show a spinner to indicate
activity.}
\usage{gtkCellRendererSpinnerNew()}
\details{Since 2.20}
\value{[\code{\link{GtkCellRenderer}}] a new \code{\link{GtkCellRenderer}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
