\alias{gtkCellRendererSpinNew}
\name{gtkCellRendererSpinNew}
\title{gtkCellRendererSpinNew}
\description{Creates a new \verb{\link{GtkCellRendererSpin}}.}
\usage{gtkCellRendererSpinNew()}
\details{  Since  2.10}
\value{[\verb{\link{GtkCellRenderer}}]  a new \verb{\link{GtkCellRendererSpin}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
