\alias{gtkCellRendererTextNew}
\name{gtkCellRendererTextNew}
\title{gtkCellRendererTextNew}
\description{Creates a new \verb{\link{GtkCellRendererText}}. Adjust how text is drawn using
object properties. Object properties can be
set globally (with \code{\link{gObjectSet}}). Also, with \verb{\link{GtkTreeViewColumn}},
you can bind a property to a value in a \verb{\link{GtkTreeModel}}. For example,
you can bind the "text" property on the cell renderer to a string
value in the model, thus rendering a different string in each row
of the \verb{\link{GtkTreeView}}}
\usage{gtkCellRendererTextNew()}
\value{[\verb{\link{GtkCellRenderer}}]  the new cell renderer}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
