\alias{gtkCellRendererToggleGetActivatable}
\name{gtkCellRendererToggleGetActivatable}
\title{gtkCellRendererToggleGetActivatable}
\description{Returns whether the cell renderer is activatable. See
\code{\link{gtkCellRendererToggleSetActivatable}}.}
\usage{gtkCellRendererToggleGetActivatable(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkCellRendererToggle}}}}
\details{Since 2.18}
\value{[logical] \code{TRUE} if the cell renderer is activatable.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
