\alias{gtkCellRendererToggleGetActive}
\name{gtkCellRendererToggleGetActive}
\title{gtkCellRendererToggleGetActive}
\description{Returns whether the cell renderer is active. See
\code{\link{gtkCellRendererToggleSetActive}}.}
\usage{gtkCellRendererToggleGetActive(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkCellRendererToggle}}]  a \verb{\link{GtkCellRendererToggle}}}}
\value{[logical] \code{TRUE} if the cell renderer is active.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
