\alias{gtkCellRendererToggleSetActivatable}
\name{gtkCellRendererToggleSetActivatable}
\title{gtkCellRendererToggleSetActivatable}
\description{Makes the cell renderer activatable.}
\usage{gtkCellRendererToggleSetActivatable(object, setting)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkCellRendererToggle}}.}
\item{\verb{setting}}{the value to set.}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
