\alias{gtkCellViewGetCellRenderers}
\name{gtkCellViewGetCellRenderers}
\title{gtkCellViewGetCellRenderers}
\description{Returns the cell renderers which have been added to \code{cell.view}.}
\usage{gtkCellViewGetCellRenderers(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkCellView}}]  a \code{\link{GtkCellView}}}}
\details{  Since  2.6}
\value{[list]  a list of cell renderers. The list.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
