\alias{gtkCellViewGetModel}
\name{gtkCellViewGetModel}
\title{gtkCellViewGetModel}
\description{Returns the model for \code{cell.view}. If no model is used \code{NULL} is
returned.}
\usage{gtkCellViewGetModel(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkCellView}}}}
\details{Since 2.16}
\value{[\code{\link{GtkTreeModel}}] a \code{\link{GtkTreeModel}} used or \code{NULL}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
