\alias{gtkCellViewNew}
\name{gtkCellViewNew}
\title{gtkCellViewNew}
\description{Creates a new \code{\link{GtkCellView}} widget.}
\usage{gtkCellViewNew(show = TRUE)}
\details{  Since  2.6}
\value{[\code{\link{GtkWidget}}]  A newly created \code{\link{GtkCellView}} widget.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
