\alias{gtkCellViewNewWithMarkup}
\name{gtkCellViewNewWithMarkup}
\title{gtkCellViewNewWithMarkup}
\description{Creates a new \verb{\link{GtkCellView}} widget, adds a \verb{\link{GtkCellRendererText}} 
to it, and makes its show \code{markup}. The text can text can be
marked up with the Pango text 
markup language.}
\usage{gtkCellViewNewWithMarkup(markup)}
\arguments{\item{\verb{markup}}{[character]  the text to display in the cell view}}
\details{  Since  2.6}
\value{[\verb{\link{GtkWidget}}]  A newly created \verb{\link{GtkCellView}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
