\alias{gtkCellViewNewWithPixbuf}
\name{gtkCellViewNewWithPixbuf}
\title{gtkCellViewNewWithPixbuf}
\description{Creates a new \verb{\link{GtkCellView}} widget, adds a \verb{\link{GtkCellRendererPixbuf}} 
to it, and makes its show \code{pixbuf}.}
\usage{gtkCellViewNewWithPixbuf(pixbuf)}
\arguments{\item{\verb{pixbuf}}{[\verb{\link{GdkPixbuf}}]  the image to display in the cell view}}
\details{  Since  2.6}
\value{[\verb{\link{GtkWidget}}]  A newly created \verb{\link{GtkCellView}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
