\alias{gtkCellViewNewWithText}
\name{gtkCellViewNewWithText}
\title{gtkCellViewNewWithText}
\description{Creates a new \verb{\link{GtkCellView}} widget, adds a \verb{\link{GtkCellRendererText}} 
to it, and makes its show \code{text}.}
\usage{gtkCellViewNewWithText(text)}
\arguments{\item{\verb{text}}{[character]  the text to display in the cell view}}
\details{  Since  2.6}
\value{[\verb{\link{GtkWidget}}]  A newly created \verb{\link{GtkCellView}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
