\alias{gtkCheckButtonNew}
\name{gtkCheckButtonNew}
\title{gtkCheckButtonNew}
\description{Creates a new \code{\link{GtkCheckButton}}.}
\usage{gtkCheckButtonNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a \code{\link{GtkWidget}}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
