\alias{gtkCheckButtonNewWithLabel}
\name{gtkCheckButtonNewWithLabel}
\title{gtkCheckButtonNewWithLabel}
\description{Creates a new \verb{\link{GtkCheckButton}} with a \verb{\link{GtkLabel}} to the right of it.}
\usage{gtkCheckButtonNewWithLabel(label, show = TRUE)}
\arguments{\item{\verb{label}}{[character] the text for the check button.}}
\value{[\verb{\link{GtkWidget}}] a \verb{\link{GtkWidget}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
