\alias{gtkCheckMenuItemGetInconsistent}
\name{gtkCheckMenuItemGetInconsistent}
\title{gtkCheckMenuItemGetInconsistent}
\description{Retrieves the value set by \code{\link{gtkCheckMenuItemSetInconsistent}}.}
\usage{gtkCheckMenuItemGetInconsistent(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkCheckMenuItem}}]  a \code{\link{GtkCheckMenuItem}}}}
\value{[logical]  \code{TRUE} if inconsistent}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
