\alias{gtkCheckMenuItemNew}
\name{gtkCheckMenuItemNew}
\title{gtkCheckMenuItemNew}
\description{Creates a new \code{\link{GtkCheckMenuItem}}.}
\usage{gtkCheckMenuItemNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkCheckMenuItem}}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
