\alias{gtkCheckMenuItemNewWithLabel}
\name{gtkCheckMenuItemNewWithLabel}
\title{gtkCheckMenuItemNewWithLabel}
\description{Creates a new \verb{\link{GtkCheckMenuItem}} with a label.}
\usage{gtkCheckMenuItemNewWithLabel(label, show = TRUE)}
\arguments{\item{\verb{label}}{[character] the string to use for the label.}}
\value{[\verb{\link{GtkWidget}}] a new \verb{\link{GtkCheckMenuItem}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
