\alias{gtkCheckMenuItemSetActive}
\name{gtkCheckMenuItemSetActive}
\title{gtkCheckMenuItemSetActive}
\description{Sets the active state of the menu item's check box.}
\usage{gtkCheckMenuItemSetActive(object, is.active)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkCheckMenuItem}}.}
\item{\verb{is.active}}{boolean value indicating whether the check box is active.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
