\alias{gtkCheckMenuItemSetDrawAsRadio}
\name{gtkCheckMenuItemSetDrawAsRadio}
\title{gtkCheckMenuItemSetDrawAsRadio}
\description{Sets whether \code{check.menu.item} is drawn like a \code{\link{GtkRadioMenuItem}}}
\usage{gtkCheckMenuItemSetDrawAsRadio(object, draw.as.radio)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCheckMenuItem}}]  a \code{\link{GtkCheckMenuItem}}}
\item{\code{draw.as.radio}}{[logical]  whether \code{check.menu.item} is drawn like a \code{\link{GtkRadioMenuItem}}}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
