\alias{gtkCheckMenuItemSetShowToggle}
\name{gtkCheckMenuItemSetShowToggle}
\title{gtkCheckMenuItemSetShowToggle}
\description{
Controls whether the check box is shown at all times.
Normally the check box is shown only when it is active or while the
menu item is selected.
\strong{WARNING: \code{gtk_check_menu_item_set_show_toggle} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCheckMenuItemSetShowToggle(object, always)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkCheckMenuItem}}.}
\item{\verb{always}}{boolean value indicating whether to always show the check box.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
