\alias{gtkCheckMenuItemToggled}
\name{gtkCheckMenuItemToggled}
\title{gtkCheckMenuItemToggled}
\description{Emits the GtkCheckMenuItem::toggled signal.}
\usage{gtkCheckMenuItemToggled(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkCheckMenuItem}}] a \code{\link{GtkCheckMenuItem}}.}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
