\alias{gtkClipboardGetDisplay}
\name{gtkClipboardGetDisplay}
\title{gtkClipboardGetDisplay}
\description{Gets the \verb{\link{GdkDisplay}} associated with \code{clipboard}}
\usage{gtkClipboardGetDisplay(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkClipboard}}]  a \verb{\link{GtkClipboard}}}}
\details{  Since  2.2}
\value{[\verb{\link{GdkDisplay}}]  the \verb{\link{GdkDisplay}} associated with \code{clipboard}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
