\alias{gtkCListAppend}
\name{gtkCListAppend}
\title{gtkCListAppend}
\description{
Adds a row to the CList at the bottom.
\strong{WARNING: \code{gtk_clist_append} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListAppend(object, text)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{text}}{An list of strings to add.}
}
\value{[integer] The number of the row added.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
