\alias{gtkCListColumnsAutosize}
\name{gtkCListColumnsAutosize}
\title{gtkCListColumnsAutosize}
\description{
Auto-sizes all columns in the CList and returns the total width of the CList.
\strong{WARNING: \code{gtk_clist_columns_autosize} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListColumnsAutosize(object)}
\arguments{\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}}
\value{[integer] The total width of the CList.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
