\alias{gtkCListColumnTitlePassive}
\name{gtkCListColumnTitlePassive}
\title{gtkCListColumnTitlePassive}
\description{
Causes the specified column title button to become passive, i.e., does
not respond to events, such as the user clicking on it.
\strong{WARNING: \code{gtk_clist_column_title_passive} is deprecated and should not be used in newly-written code.  
Causes the specified column title button to become passive, i.e., does
not respond to events, such as the user clicking on it.}
}
\usage{gtkCListColumnTitlePassive(object, column)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCList}}] The \code{\link{GtkCList}} to affect.}
\item{\code{column}}{[integer] The column to make passive, counting from 0.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
