\alias{gtkCListColumnTitlesActive}
\name{gtkCListColumnTitlesActive}
\title{gtkCListColumnTitlesActive}
\description{
Causes all column title buttons to become active.  This is the same
as calling \code{\link{gtkCListColumnTitleActive}} for each column.
\strong{WARNING: \code{gtk_clist_column_titles_active} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListColumnTitlesActive(object)}
\arguments{\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
