\alias{gtkCListColumnTitlesHide}
\name{gtkCListColumnTitlesHide}
\title{gtkCListColumnTitlesHide}
\description{
Causes the \verb{\link{GtkCList}} to hide its column titles, if they are currently
showing.
\strong{WARNING: \code{gtk_clist_column_titles_hide} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListColumnTitlesHide(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkCList}}] The \verb{\link{GtkCList}} to affect.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
