\alias{gtkCListColumnTitlesPassive}
\name{gtkCListColumnTitlesPassive}
\title{gtkCListColumnTitlesPassive}
\description{
Causes all column title buttons to become passive.  This is the same
as calling \code{\link{gtkCListColumnTitlePassive}} for each column.
\strong{WARNING: \code{gtk_clist_column_titles_passive} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListColumnTitlesPassive(object)}
\arguments{\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
