\alias{gtkCListColumnTitlesShow}
\name{gtkCListColumnTitlesShow}
\title{gtkCListColumnTitlesShow}
\description{
This function causes the \code{\link{GtkCList}} to show its column titles, if
they are not already showing.
\strong{WARNING: \code{gtk_clist_column_titles_show} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListColumnTitlesShow(object)}
\arguments{\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
