\alias{gtkCListGetHadjustment}
\name{gtkCListGetHadjustment}
\title{gtkCListGetHadjustment}
\description{
Gets the \verb{\link{GtkAdjustment}} currently being used for the horizontal
aspect.
\strong{WARNING: \code{gtk_clist_get_hadjustment} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListGetHadjustment(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkCList}}] The \verb{\link{GtkCList}} to check.}}
\value{[\verb{\link{GtkAdjustment}}] A \verb{\link{GtkAdjustment}} object, or NULL if none is currently
being used.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
