\alias{gtkCListGetRowData}
\name{gtkCListGetRowData}
\title{gtkCListGetRowData}
\description{
Gets the currently set data for the specified row.
\strong{WARNING: \code{gtk_clist_get_row_data} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListGetRowData(object, row)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCList}}] The \code{\link{GtkCList}} to affect.}
\item{\code{row}}{[integer] The row to query.}
}
\value{[R object] The data set for the row.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
