\alias{gtkCListGetSelectable}
\name{gtkCListGetSelectable}
\title{gtkCListGetSelectable}
\description{
Gets whether the specified row is selectable or not.
\strong{WARNING: \code{gtk_clist_get_selectable} is deprecated and should not be used in newly-written code.  
Gets whether the specified row is selectable or not.}
}
\usage{gtkCListGetSelectable(object, row)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCList}}] The \code{\link{GtkCList}} to affect.}
\item{\code{row}}{[integer] The row to query.}
}
\value{[logical] A \code{logical} value.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
