\alias{gtkCListOptimalColumnWidth}
\name{gtkCListOptimalColumnWidth}
\title{gtkCListOptimalColumnWidth}
\description{
Gets the required width in pixels that is needed to show
everything in the specified column.
\strong{WARNING: \code{gtk_clist_optimal_column_width} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListOptimalColumnWidth(object, column)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCList}}] The \verb{\link{GtkCList}} to check.}
\item{\verb{column}}{[integer] The column to check.}
}
\value{[integer] The required width in pixels for the column.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
