\alias{gtkCListRowIsVisible}
\name{gtkCListRowIsVisible}
\title{gtkCListRowIsVisible}
\description{
Checks how the specified row is visible.
\strong{WARNING: \code{gtk_clist_row_is_visible} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListRowIsVisible(object, row)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{row}}{The row to query.}
}
\value{[\code{\link{GtkVisibility}}] A \code{\link{GtkVisibility}} value that tells you how the row is visible.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
