\alias{gtkCListSelectAll}
\name{gtkCListSelectAll}
\title{gtkCListSelectAll}
\description{
Selects all rows in the CList.  This function has no affect for a
CList in "single" or "browse" selection mode.
\strong{WARNING: \code{gtk_clist_select_all} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSelectAll(object)}
\arguments{\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
