\alias{gtkCListSelectRow}
\name{gtkCListSelectRow}
\title{gtkCListSelectRow}
\description{
Selects the specified row.  Causes the "select-row" signal to be emitted for the specified row and column.
\strong{WARNING: \code{gtk_clist_select_row} is deprecated and should not be used in newly-written code.  
Selects the specified row.  Causes the "select-row" signal to be emitted for the specified row and column.}
}
\usage{gtkCListSelectRow(object, row, column)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCList}}] The \code{\link{GtkCList}} to affect.}
\item{\code{row}}{[integer] The row to select.}
\item{\code{column}}{[integer] The column to select.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
