\alias{gtkCListSetCellStyle}
\name{gtkCListSetCellStyle}
\title{gtkCListSetCellStyle}
\description{
Sets the style for the specified cell.
\strong{WARNING: \code{gtk_clist_set_cell_style} is deprecated and should not be used in newly-written code.  
Sets the style for the specified cell.}
}
\usage{gtkCListSetCellStyle(object, row, column, style)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCList}}] The \code{\link{GtkCList}} to affect.}
\item{\code{row}}{[integer] The row of the cell.}
\item{\code{column}}{[integer] The column of the cell.}
\item{\code{style}}{[\code{\link{GtkStyle}}] A pointer to a \code{\link{GtkStyle}} structure.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
