\alias{gtkCListSetForeground}
\name{gtkCListSetForeground}
\title{gtkCListSetForeground}
\description{
Sets the foreground color for the specified row.
\strong{WARNING: \code{gtk_clist_set_foreground} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetForeground(object, row, color)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCList}}] The \verb{\link{GtkCList}} to affect.}
\item{\verb{row}}{[integer] The row to affect.}
\item{\verb{color}}{[\verb{\link{GdkColor}}] A pointer to a \verb{\link{GdkColor}} structure.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
