\alias{gtkCListSetHadjustment}
\name{gtkCListSetHadjustment}
\title{gtkCListSetHadjustment}
\description{
Allows you to set the \verb{\link{GtkAdjustment}} to be used for the horizontal
aspect of the \verb{\link{GtkCList}} widget.
\strong{WARNING: \code{gtk_clist_set_hadjustment} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetHadjustment(object, adjustment)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCList}}] The \verb{\link{GtkCList}} to affect.}
\item{\verb{adjustment}}{[\verb{\link{GtkAdjustment}}] A pointer to a \verb{\link{GtkAdjustment}} widget, or NULL.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
