\alias{gtkCListSetReorderable}
\name{gtkCListSetReorderable}
\title{gtkCListSetReorderable}
\description{
Sets whether the CList's rows are re-orderable using drag-and-drop.
\strong{WARNING: \code{gtk_clist_set_reorderable} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetReorderable(object, reorderable)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{reorderable}}{\code{TRUE} or \code{FALSE}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
