\alias{gtkCListSetRowDataFull}
\name{gtkCListSetRowDataFull}
\title{gtkCListSetRowDataFull}
\description{
Sets the data for specified row, with a callback when the row is destroyed.
\strong{WARNING: \code{gtk_clist_set_row_data_full} is deprecated and should not be used in newly-written code.  
Sets the data for specified row, with a callback when the row is destroyed.}
}
\usage{gtkCListSetRowDataFull(object, row, data = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCList}}] The \code{\link{GtkCList}} to affect.}
\item{\code{row}}{[integer] The row to affect.}
\item{\code{data}}{[R object] The data to set for the row.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
