\alias{gtkCListSetRowHeight}
\name{gtkCListSetRowHeight}
\title{gtkCListSetRowHeight}
\description{
Causes the \verb{\link{GtkCList}} to have a specified height for its 
rows.  Setting the row height to 0 allows the \verb{\link{GtkCList}} to adjust
automatically to data in the row.
\strong{WARNING: \code{gtk_clist_set_row_height} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetRowHeight(object, height)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCList}}] The \verb{\link{GtkCList}} to affect.}
\item{\verb{height}}{[numeric] The height, in pixels.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
