\alias{gtkCListSetRowStyle}
\name{gtkCListSetRowStyle}
\title{gtkCListSetRowStyle}
\description{
Sets the style for all cells in the specified row.
\strong{WARNING: \code{gtk_clist_set_row_style} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetRowStyle(object, row, style)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCList}}] The \verb{\link{GtkCList}} to affect.}
\item{\verb{row}}{[integer] The row to affect.}
\item{\verb{style}}{[\verb{\link{GtkStyle}}] A pointer to a \verb{\link{GtkStyle}} to set.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
