\alias{gtkCListSetShadowType}
\name{gtkCListSetShadowType}
\title{gtkCListSetShadowType}
\description{
Sets the shadow type for the specified CList.  Changing this value
will cause the \verb{\link{GtkCList}} to update its visuals.
\strong{WARNING: \code{gtk_clist_set_shadow_type} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetShadowType(object, type)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCList}}] The \verb{\link{GtkCList}} to affect.}
\item{\verb{type}}{[\code{\link{GtkShadowType}}] The GtkShadowType desired.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
