\alias{gtkCListSetSortColumn}
\name{gtkCListSetSortColumn}
\title{gtkCListSetSortColumn}
\description{
Sets the sort column of the clist. The sort column is used by the
default compare function to determine which column to sort by.
\strong{WARNING: \code{gtk_clist_set_sort_column} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetSortColumn(object, column)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{column}}{The column to sort by}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
