\alias{gtkCListSetSortType}
\name{gtkCListSetSortType}
\title{gtkCListSetSortType}
\description{
Sets the sort type of the \code{GtkClist}. This is either GTK\_SORT\_ASCENDING for
ascening sort or GTK\_SORT\_DESCENDING for descending sort.
\strong{WARNING: \code{gtk_clist_set_sort_type} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetSortType(object, sort.type)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCList}}] The \code{\link{GtkCList}} to affect.}
\item{\code{sort.type}}{[\code{\link{GtkSortType}}] the \code{\link{GtkSortType}} to use}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
