\alias{gtkCListSetUseDragIcons}
\name{gtkCListSetUseDragIcons}
\title{gtkCListSetUseDragIcons}
\description{
Determines whether the \code{GtkClist} should use icons when
doing drag-and-drop operations.
\strong{WARNING: \code{gtk_clist_set_use_drag_icons} is deprecated and should not be used in newly-written code.  
Determines whether the \code{GtkClist} should use icons when
doing drag-and-drop operations.}
}
\usage{gtkCListSetUseDragIcons(object, use.icons)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCList}}] The \code{\link{GtkCList}} to affect.}
\item{\code{use.icons}}{[logical] \code{TRUE} or \code{FALSE}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
