\alias{gtkCListSwapRows}
\name{gtkCListSwapRows}
\title{gtkCListSwapRows}
\description{
Swaps the two specified rows with each other.
\strong{WARNING: \code{gtk_clist_swap_rows} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSwapRows(object, row1, row2)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCList}}] The \code{\link{GtkCList}} to affect.}
\item{\code{row1}}{[integer] Number of the first row.}
\item{\code{row2}}{[integer] Number of the second row.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
