\alias{gtkCListThaw}
\name{gtkCListThaw}
\title{gtkCListThaw}
\description{
Causes the specified \code{\link{GtkCList}} to allow visual updates.
\strong{WARNING: \code{gtk_clist_thaw} is deprecated and should not be used in newly-written code.  
Causes the specified \code{\link{GtkCList}} to allow visual updates.}
}
\usage{gtkCListThaw(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkCList}}] The \code{\link{GtkCList}} to thaw.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
