\alias{gtkCListUndoSelection}
\name{gtkCListUndoSelection}
\title{gtkCListUndoSelection}
\description{
Undoes the last selection for an "extended selection mode" CList.
\strong{WARNING: \code{gtk_clist_undo_selection} is deprecated and should not be used in newly-written code.  
Undoes the last selection for an "extended selection mode" CList.}
}
\usage{gtkCListUndoSelection(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkCList}}] The \code{\link{GtkCList}} to affect.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
