\alias{gtkCListUnselectAll}
\name{gtkCListUnselectAll}
\title{gtkCListUnselectAll}
\description{
Unselects all rows in the CList.
\strong{WARNING: \code{gtk_clist_unselect_all} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListUnselectAll(object)}
\arguments{\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
