\alias{gtkColorButtonGetAlpha}
\name{gtkColorButtonGetAlpha}
\title{gtkColorButtonGetAlpha}
\description{Returns the current alpha value.}
\usage{gtkColorButtonGetAlpha(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkColorButton}}.}}
\details{Since 2.4}
\value{[integer] an integer between 0 and 65535.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
