\alias{gtkColorButtonGetColor}
\name{gtkColorButtonGetColor}
\title{gtkColorButtonGetColor}
\description{Sets \code{color} to be the current color in the \verb{\link{GtkColorButton}} widget.}
\usage{gtkColorButtonGetColor(object, color)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkColorButton}}]  a \verb{\link{GtkColorButton}}.}}
\details{  Since  2.4}
\value{
A list containing the following elements:
\item{\verb{color}}{[\verb{\link{GdkColor}}]  a \verb{\link{GdkColor}} to fill in with the current color.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
