\alias{gtkColorButtonGetTitle}
\name{gtkColorButtonGetTitle}
\title{gtkColorButtonGetTitle}
\description{Gets the title of the color selection dialog.}
\usage{gtkColorButtonGetTitle(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkColorButton}}}}
\details{Since 2.4}
\value{[character] An internal string, do not free the return value}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
