\alias{gtkColorButtonGetUseAlpha}
\name{gtkColorButtonGetUseAlpha}
\title{gtkColorButtonGetUseAlpha}
\description{Does the color selection dialog use the alpha channel?}
\usage{gtkColorButtonGetUseAlpha(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkColorButton}}.}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if the color sample uses alpha channel, \code{FALSE} if not.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
