\alias{gtkColorButtonNew}
\name{gtkColorButtonNew}
\title{gtkColorButtonNew}
\description{Creates a new color button. This returns a widget in the form of
a small button containing a swatch representing the current selected 
color. When the button is clicked, a color-selection dialog will open, 
allowing the user to select a color. The swatch will be updated to reflect 
the new color when the user finishes.}
\usage{gtkColorButtonNew(show = TRUE)}
\details{Since 2.4}
\value{[\code{\link{GtkWidget}}] a new color button.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
