\alias{gtkColorButtonNewWithColor}
\name{gtkColorButtonNewWithColor}
\title{gtkColorButtonNewWithColor}
\description{Creates a new color button.}
\usage{gtkColorButtonNewWithColor(color, show = TRUE)}
\arguments{\item{\verb{color}}{A \code{\link{GdkColor}} to set the current color with.}}
\details{Since 2.4}
\value{[\code{\link{GtkWidget}}] a new color button.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
