\alias{gtkColorButtonSetAlpha}
\name{gtkColorButtonSetAlpha}
\title{gtkColorButtonSetAlpha}
\description{Sets the current opacity to be \code{alpha}.}
\usage{gtkColorButtonSetAlpha(object, alpha)}
\arguments{
\item{\code{object}}{[\code{\link{GtkColorButton}}]  a \code{\link{GtkColorButton}}.}
\item{\code{alpha}}{[integer]  an integer between 0 and 65535.}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
