\alias{gtkColorButtonSetColor}
\name{gtkColorButtonSetColor}
\title{gtkColorButtonSetColor}
\description{Sets the current color to be \code{color}.}
\usage{gtkColorButtonSetColor(object, color)}
\arguments{
\item{\code{object}}{[\code{\link{GtkColorButton}}]  a \code{\link{GtkColorButton}}.}
\item{\code{color}}{[\code{\link{GdkColor}}]  A \code{\link{GdkColor}} to set the current color with.}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
