\alias{gtkColorButtonSetTitle}
\name{gtkColorButtonSetTitle}
\title{gtkColorButtonSetTitle}
\description{Sets the title for the color selection dialog.}
\usage{gtkColorButtonSetTitle(object, title)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkColorButton}}]  a \verb{\link{GtkColorButton}}}
\item{\verb{title}}{[character]  String containing new window title.}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
