\alias{gtkColorButtonSetUseAlpha}
\name{gtkColorButtonSetUseAlpha}
\title{gtkColorButtonSetUseAlpha}
\description{Sets whether or not the color button should use the alpha channel.}
\usage{gtkColorButtonSetUseAlpha(object, use.alpha)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkColorButton}}]  a \verb{\link{GtkColorButton}}.}
\item{\verb{use.alpha}}{[logical] \code{TRUE} if color button should use alpha channel, \code{FALSE} if not.}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
