\alias{gtkColorSelectionDialogGetColorSelection}
\name{gtkColorSelectionDialogGetColorSelection}
\title{gtkColorSelectionDialogGetColorSelection}
\description{Retrieves the \code{\link{GtkColorSelection}} widget embedded in the dialog.}
\usage{gtkColorSelectionDialogGetColorSelection(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkColorSelectionDialog}}}}
\details{Since 2.14}
\value{[\code{\link{GtkWidget}}] the embedded \code{\link{GtkColorSelection}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
