\alias{gtkColorSelectionDialogNew}
\name{gtkColorSelectionDialogNew}
\title{gtkColorSelectionDialogNew}
\description{Creates a new \verb{\link{GtkColorSelectionDialog}}.}
\usage{gtkColorSelectionDialogNew(title = NULL, show = TRUE)}
\arguments{\item{\verb{title}}{[character] a string containing the title text for the dialog.}}
\value{[\verb{\link{GtkWidget}}] a \verb{\link{GtkColorSelectionDialog}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
