\alias{gtkColorSelectionGetCurrentAlpha}
\name{gtkColorSelectionGetCurrentAlpha}
\title{gtkColorSelectionGetCurrentAlpha}
\description{Returns the current alpha value.}
\usage{gtkColorSelectionGetCurrentAlpha(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkColorSelection}}.}}
\value{[integer] an integer between 0 and 65535.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
