\alias{gtkColorSelectionGetCurrentColor}
\name{gtkColorSelectionGetCurrentColor}
\title{gtkColorSelectionGetCurrentColor}
\description{Sets \code{color} to be the current color in the GtkColorSelection widget.}
\usage{gtkColorSelectionGetCurrentColor(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkColorSelection}}]  a \verb{\link{GtkColorSelection}}.}}
\value{
A list containing the following elements:
\item{\verb{color}}{[\verb{\link{GdkColor}}]  a \verb{\link{GdkColor}} to fill in with the current color.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
