\alias{gtkColorSelectionGetHasOpacityControl}
\name{gtkColorSelectionGetHasOpacityControl}
\title{gtkColorSelectionGetHasOpacityControl}
\description{Determines whether the colorsel has an opacity control.}
\usage{gtkColorSelectionGetHasOpacityControl(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkColorSelection}}]  a \code{\link{GtkColorSelection}}.}}
\value{[logical]  \code{TRUE} if the \code{colorsel} has an opacity control.  \code{FALSE} if it does't.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
