\alias{gtkColorSelectionGetHasPalette}
\name{gtkColorSelectionGetHasPalette}
\title{gtkColorSelectionGetHasPalette}
\description{Determines whether the color selector has a color palette.}
\usage{gtkColorSelectionGetHasPalette(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkColorSelection}}]  a \code{\link{GtkColorSelection}}.}}
\value{[logical]  \code{TRUE} if the selector has a palette.  \code{FALSE} if it hasn't.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
