\alias{gtkColorSelectionGetPreviousAlpha}
\name{gtkColorSelectionGetPreviousAlpha}
\title{gtkColorSelectionGetPreviousAlpha}
\description{Returns the previous alpha value.}
\usage{gtkColorSelectionGetPreviousAlpha(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkColorSelection}}.}}
\value{[integer] an integer between 0 and 65535.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
