\alias{gtkColorSelectionGetPreviousColor}
\name{gtkColorSelectionGetPreviousColor}
\title{gtkColorSelectionGetPreviousColor}
\description{Fills \code{color} in with the original color value.}
\usage{gtkColorSelectionGetPreviousColor(object, color)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkColorSelection}}]  a \verb{\link{GtkColorSelection}}.}
\item{\verb{color}}{[\verb{\link{GdkColor}}]  a \verb{\link{GdkColor}} to fill in with the original color value.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
