\alias{gtkColorSelectionNew}
\name{gtkColorSelectionNew}
\title{gtkColorSelectionNew}
\description{Creates a new GtkColorSelection.}
\usage{gtkColorSelectionNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}]  a new \code{\link{GtkColorSelection}}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
