\alias{gtkColorSelectionPaletteToString}
\name{gtkColorSelectionPaletteToString}
\title{gtkColorSelectionPaletteToString}
\description{Encodes a palette as a string, useful for persistent storage.}
\usage{gtkColorSelectionPaletteToString(colors)}
\arguments{\item{\verb{colors}}{a list of colors.}}
\value{[character] allocated string encoding the palette.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
