\alias{gtkColorSelectionSetCurrentAlpha}
\name{gtkColorSelectionSetCurrentAlpha}
\title{gtkColorSelectionSetCurrentAlpha}
\description{Sets the current opacity to be \code{alpha}.  The first time this is called, it will
also set the original opacity to be \code{alpha} too.}
\usage{gtkColorSelectionSetCurrentAlpha(object, alpha)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkColorSelection}}.}
\item{\verb{alpha}}{an integer between 0 and 65535.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
