\alias{gtkColorSelectionSetPreviousAlpha}
\name{gtkColorSelectionSetPreviousAlpha}
\title{gtkColorSelectionSetPreviousAlpha}
\description{Sets the 'previous' alpha to be \code{alpha}.  This function should be called with
some hesitations, as it might seem confusing to have that alpha change.}
\usage{gtkColorSelectionSetPreviousAlpha(object, alpha)}
\arguments{
\item{\code{object}}{[\code{\link{GtkColorSelection}}]  a \code{\link{GtkColorSelection}}.}
\item{\code{alpha}}{[integer]  an integer between 0 and 65535.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
