\alias{gtkComboBoxEntryGetTextColumn}
\name{gtkComboBoxEntryGetTextColumn}
\title{gtkComboBoxEntryGetTextColumn}
\description{Returns the column which \code{entry.box} is using to get the strings from.}
\usage{gtkComboBoxEntryGetTextColumn(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkComboBoxEntry}}]  A \verb{\link{GtkComboBoxEntry}}.}}
\details{  Since  2.4}
\value{[integer]  A column in the data source model of \code{entry.box}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
