\alias{gtkComboBoxEntryNewText}
\name{gtkComboBoxEntryNewText}
\title{gtkComboBoxEntryNewText}
\description{Convenience function which constructs a new editable text combo box, which 
is a \verb{\link{GtkComboBoxEntry}} just displaying strings. If you use this function to
create a text combo box, you should only manipulate its data source with
the following convenience functions: \code{\link{gtkComboBoxAppendText}},
\code{\link{gtkComboBoxInsertText}}, \code{\link{gtkComboBoxPrependText}} and
\code{\link{gtkComboBoxRemoveText}}.}
\usage{gtkComboBoxEntryNewText()}
\details{  Since  2.4}
\value{[\verb{\link{GtkWidget}}]  A new text \verb{\link{GtkComboBoxEntry}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
