\alias{gtkComboBoxGetAddTearoffs}
\name{gtkComboBoxGetAddTearoffs}
\title{gtkComboBoxGetAddTearoffs}
\description{Gets the current value of the :add-tearoffs property.}
\usage{gtkComboBoxGetAddTearoffs(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkComboBox}}}}
\value{[logical] the current value of the :add-tearoffs property.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
