\alias{gtkComboBoxGetColumnSpanColumn}
\name{gtkComboBoxGetColumnSpanColumn}
\title{gtkComboBoxGetColumnSpanColumn}
\description{Returns the column with column span information for \code{combo.box}.}
\usage{gtkComboBoxGetColumnSpanColumn(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkComboBox}}}}
\details{Since 2.6}
\value{[integer] the column span column.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
