\alias{gtkComboBoxGetFocusOnClick}
\name{gtkComboBoxGetFocusOnClick}
\title{gtkComboBoxGetFocusOnClick}
\description{Returns whether the combo box grabs focus when it is clicked 
with the mouse. See \code{\link{gtkComboBoxSetFocusOnClick}}.}
\usage{gtkComboBoxGetFocusOnClick(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkComboBox}}}}
\details{Since 2.6}
\value{[logical] \code{TRUE} if the combo box grabs focus when it is
clicked with the mouse.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
