\alias{gtkComboBoxGetModel}
\name{gtkComboBoxGetModel}
\title{gtkComboBoxGetModel}
\description{Returns the \verb{\link{GtkTreeModel}} which is acting as data source for \code{combo.box}.}
\usage{gtkComboBoxGetModel(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkComboBox}}]  A \verb{\link{GtkComboBox}}}}
\details{  Since  2.4}
\value{[\verb{\link{GtkTreeModel}}]  A \verb{\link{GtkTreeModel}} which was passed during construction.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
