\alias{gtkComboBoxGetPopupAccessible}
\name{gtkComboBoxGetPopupAccessible}
\title{gtkComboBoxGetPopupAccessible}
\description{Gets the accessible object corresponding to the combo box's popup.}
\usage{gtkComboBoxGetPopupAccessible(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkComboBox}}}}
\details{This function is mostly intended for use by accessibility technologies;
applications should have little use for it.
  Since 2.6}
\value{[\code{\link{AtkObject}}] the accessible object corresponding to the combo box's popup.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
