\alias{gtkComboBoxGetRowSeparatorFunc}
\name{gtkComboBoxGetRowSeparatorFunc}
\title{gtkComboBoxGetRowSeparatorFunc}
\description{Returns the current row separator function.}
\usage{gtkComboBoxGetRowSeparatorFunc(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkComboBox}}]  a \verb{\link{GtkComboBox}}}}
\details{  Since  2.6}
\value{[\verb{\link{GtkTreeViewRowSeparatorFunc}}]  the current row separator function.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
