\alias{gtkComboBoxGetRowSpanColumn}
\name{gtkComboBoxGetRowSpanColumn}
\title{gtkComboBoxGetRowSpanColumn}
\description{Returns the column with row span information for \code{combo.box}.}
\usage{gtkComboBoxGetRowSpanColumn(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkComboBox}}]  A \code{\link{GtkComboBox}}}}
\details{  Since  2.6}
\value{[integer]  the row span column.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
