\alias{gtkComboBoxGetTitle}
\name{gtkComboBoxGetTitle}
\title{gtkComboBoxGetTitle}
\description{Gets the current title of the menu in tearoff mode. See
\code{\link{gtkComboBoxSetAddTearoffs}}.}
\usage{gtkComboBoxGetTitle(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkComboBox}}}}
\details{Since 2.10}
\value{[character] the menu's title in tearoff mode. This is an internal copy of the
string which must not be freed.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
