\alias{gtkComboBoxGetWrapWidth}
\name{gtkComboBoxGetWrapWidth}
\title{gtkComboBoxGetWrapWidth}
\description{Returns the wrap width which is used to determine the number of columns 
for the popup menu. If the wrap width is larger than 1, the combo box 
is in table mode.}
\usage{gtkComboBoxGetWrapWidth(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkComboBox}}}}
\details{Since 2.6}
\value{[integer] the wrap width.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
