\alias{gtkComboBoxNew}
\name{gtkComboBoxNew}
\title{gtkComboBoxNew}
\description{Creates a new empty \verb{\link{GtkComboBox}}.}
\usage{gtkComboBoxNew(show = TRUE)}
\details{  Since  2.4}
\value{[\verb{\link{GtkWidget}}]  A new \verb{\link{GtkComboBox}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
