\alias{gtkComboBoxNewWithModel}
\name{gtkComboBoxNewWithModel}
\title{gtkComboBoxNewWithModel}
\description{Creates a new \verb{\link{GtkComboBox}} with the model initialized to \code{model}.}
\usage{gtkComboBoxNewWithModel(model, show = TRUE)}
\arguments{\item{\verb{model}}{[\verb{\link{GtkTreeModel}}]  A \verb{\link{GtkTreeModel}}.}}
\details{  Since  2.4}
\value{[\verb{\link{GtkWidget}}]  A new \verb{\link{GtkComboBox}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
