\alias{gtkComboBoxPopup}
\name{gtkComboBoxPopup}
\title{gtkComboBoxPopup}
\description{Pops up the menu or dropdown list of \code{combo.box}. }
\usage{gtkComboBoxPopup(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkComboBox}}}}
\details{This function is mostly intended for use by accessibility technologies;
applications should have little use for it.
  Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
