\alias{gtkComboBoxPrependText}
\name{gtkComboBoxPrependText}
\title{gtkComboBoxPrependText}
\description{Prepends \code{string} to the list of strings stored in \code{combo.box}. Note that
you can only use this function with combo boxes constructed with
\code{\link{gtkComboBoxNewText}}.}
\usage{gtkComboBoxPrependText(object, text)}
\arguments{
\item{\code{object}}{[\code{\link{GtkComboBox}}]  A \code{\link{GtkComboBox}} constructed with \code{\link{gtkComboBoxNewText}}.}
\item{\code{text}}{[character]  A string.}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
