\alias{gtkComboBoxRemoveText}
\name{gtkComboBoxRemoveText}
\title{gtkComboBoxRemoveText}
\description{Removes the string at \code{position} from \code{combo.box}. Note that you can only use
this function with combo boxes constructed with \code{\link{gtkComboBoxNewText}}.}
\usage{gtkComboBoxRemoveText(object, position)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkComboBox}}]  A \verb{\link{GtkComboBox}} constructed with \code{\link{gtkComboBoxNewText}}}
\item{\verb{position}}{[integer]  Index of the item to remove}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
