\alias{gtkComboBoxSetAddTearoffs}
\name{gtkComboBoxSetAddTearoffs}
\title{gtkComboBoxSetAddTearoffs}
\description{Sets whether the popup menu should have a tearoff 
menu item.}
\usage{gtkComboBoxSetAddTearoffs(object, add.tearoffs)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkComboBox}}]  a \verb{\link{GtkComboBox}} }
\item{\verb{add.tearoffs}}{[logical] \code{TRUE} to add tearoff menu items}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
