\alias{gtkComboBoxSetButtonSensitivity}
\name{gtkComboBoxSetButtonSensitivity}
\title{gtkComboBoxSetButtonSensitivity}
\description{Sets whether the dropdown button of the combo box should be
always sensitive (\code{GTK_SENSITIVITY_ON}), never sensitive (\code{GTK_SENSITIVITY_OFF})
or only if there is at least one item to display (\code{GTK_SENSITIVITY_AUTO}).}
\usage{gtkComboBoxSetButtonSensitivity(object, sensitivity)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkComboBox}}}
\item{\verb{sensitivity}}{specify the sensitivity of the dropdown button}
}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
