\alias{gtkComboBoxSetColumnSpanColumn}
\name{gtkComboBoxSetColumnSpanColumn}
\title{gtkComboBoxSetColumnSpanColumn}
\description{Sets the column with column span information for \code{combo.box} to be
\code{column.span}. The column span column contains integers which indicate
how many columns an item should span.}
\usage{gtkComboBoxSetColumnSpanColumn(object, column.span)}
\arguments{
\item{\code{object}}{[\code{\link{GtkComboBox}}]  A \code{\link{GtkComboBox}}}
\item{\code{column.span}}{[integer]  A column in the model passed during construction}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
