\alias{gtkComboBoxSetTitle}
\name{gtkComboBoxSetTitle}
\title{gtkComboBoxSetTitle}
\description{Sets the menu's title in tearoff mode.}
\usage{gtkComboBoxSetTitle(object, title)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkComboBox}}}
\item{\verb{title}}{a title for the menu in tearoff mode}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
