\alias{gtkComboBoxSetWrapWidth}
\name{gtkComboBoxSetWrapWidth}
\title{gtkComboBoxSetWrapWidth}
\description{Sets the wrap width of \code{combo.box} to be \code{width}. The wrap width is basically
the preferred number of columns when you want the popup to be layed out
in a table.}
\usage{gtkComboBoxSetWrapWidth(object, width)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkComboBox}}]  A \verb{\link{GtkComboBox}}}
\item{\verb{width}}{[integer]  Preferred number of columns}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
