\alias{gtkComboSetPopdownStrings}
\name{gtkComboSetPopdownStrings}
\title{gtkComboSetPopdownStrings}
\description{
Convenience function to set all of the items in the popup list.
(See the example above.)
\strong{WARNING: \code{gtk_combo_set_popdown_strings} has been deprecated since version 2.4 and should not be used in newly-written code. Use \code{\link{GtkComboBox}} instead.}
}
\usage{gtkComboSetPopdownStrings(object, strings)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkCombo}}.}
\item{\verb{strings}}{a list of strings, or \code{NULL} to clear the popup list}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
