\alias{gtkComboSetUseArrows}
\name{gtkComboSetUseArrows}
\title{gtkComboSetUseArrows}
\description{
Specifies if the arrow (cursor) keys can be used to step through the items in
the list. This is on by default.
\strong{WARNING: \code{gtk_combo_set_use_arrows} is deprecated and should not be used in newly-written code. Use \code{\link{GtkComboBox}} instead.}
}
\usage{gtkComboSetUseArrows(object, val)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCombo}}] a \code{\link{GtkCombo}}.}
\item{\code{val}}{[logical] \code{TRUE} if the arrow keys can be used to step through the items in 
     the list.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
