\alias{gtkComboSetUseArrowsAlways}
\name{gtkComboSetUseArrowsAlways}
\title{gtkComboSetUseArrowsAlways}
\description{
Obsolete function, does nothing.
\strong{WARNING: \code{gtk_combo_set_use_arrows_always} is deprecated and should not be used in newly-written code. Use \code{\link{GtkComboBox}} instead.}
}
\usage{gtkComboSetUseArrowsAlways(object, val)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCombo}}] a \code{\link{GtkCombo}}.}
\item{\code{val}}{[logical] unused}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
