\alias{gtkContainerChildGet}
\name{gtkContainerChildGet}
\title{gtkContainerChildGet}
\description{Gets the values of one or more child properties for \code{child} and \code{container}.}
\usage{gtkContainerChildGet(object, child, ...)}
\arguments{\item{\code{...}}{[\code{\link{GtkWidget}}]  a widget which is a child of \code{container}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
