\alias{gtkContainerChildSet}
\name{gtkContainerChildSet}
\title{gtkContainerChildSet}
\description{Sets one or more child properties for \code{child} and \code{container}.}
\usage{gtkContainerChildSet(object, child, ...)}
\arguments{
\item{\code{object}}{[\code{\link{GtkContainer}}]  a \code{\link{GtkContainer}}}
\item{\code{child}}{[\code{\link{GtkWidget}}]  a widget which is a child of \code{container}}
\item{\code{...}}{ a list of property names and values, starting
          with \code{first.prop.name}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
