\alias{gtkContainerChildSetProperty}
\name{gtkContainerChildSetProperty}
\title{gtkContainerChildSetProperty}
\description{Sets a child property for \code{child} and \code{container}.}
\usage{gtkContainerChildSetProperty(object, child, property.name, value)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkContainer}}]  a \verb{\link{GtkContainer}}}
\item{\verb{child}}{[\verb{\link{GtkWidget}}]  a widget which is a child of \code{container}}
\item{\verb{property.name}}{[character]  the name of the property to set}
\item{\verb{value}}{[R object]  the value to set the property to}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
