\alias{gtkContainerClassInstallChildProperty}
\name{gtkContainerClassInstallChildProperty}
\title{gtkContainerClassInstallChildProperty}
\description{Installs a child property on a container class.}
\usage{gtkContainerClassInstallChildProperty(cclass, property.id, pspec)}
\arguments{
\item{\code{cclass}}{[\code{\link{GtkContainerClass}}]  a \code{\link{GtkContainerClass}}}
\item{\code{property.id}}{[numeric]  the id for the property}
\item{\code{pspec}}{[\code{\link{GParamSpec}}]  the \code{\link{GParamSpec}} for the property}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
