\alias{gtkContainerGetBorderWidth}
\name{gtkContainerGetBorderWidth}
\title{gtkContainerGetBorderWidth}
\description{Retrieves the border width of the container. See
\code{\link{gtkContainerSetBorderWidth}}.}
\usage{gtkContainerGetBorderWidth(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkContainer}}]  a \code{\link{GtkContainer}}}}
\value{[numeric]  the current border width}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
