\alias{gtkContainerGetFocusChild}
\name{gtkContainerGetFocusChild}
\title{gtkContainerGetFocusChild}
\description{Returns the current focus child widget inside \code{container}.}
\usage{gtkContainerGetFocusChild(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkContainer}}}}
\details{Since 2.14}
\value{[\code{\link{GtkWidget}}] The child widget which has the focus
inside \code{container}, or \code{NULL} if none is set.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
