\alias{gtkContainerGetFocusVadjustment}
\name{gtkContainerGetFocusVadjustment}
\title{gtkContainerGetFocusVadjustment}
\description{Retrieves the vertical focus adjustment for the container. See
\code{\link{gtkContainerSetFocusVadjustment}}.}
\usage{gtkContainerGetFocusVadjustment(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkContainer}}]  a \code{\link{GtkContainer}}}}
\value{[\code{\link{GtkAdjustment}}]  the vertical focus adjustment, or \code{NULL} if
  none has been set.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
