\alias{gtkContainerGetResizeMode}
\name{gtkContainerGetResizeMode}
\title{gtkContainerGetResizeMode}
\description{Returns the resize mode for the container. See
\code{\link{gtkContainerSetResizeMode}}.}
\usage{gtkContainerGetResizeMode(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkContainer}}}}
\value{[\code{\link{GtkResizeMode}}] the current resize mode}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
