\alias{gtkContainerResizeChildren}
\name{gtkContainerResizeChildren}
\title{gtkContainerResizeChildren}
\description{\emph{undocumented
}}
\usage{gtkContainerResizeChildren(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkContainer}}] }}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
