\alias{gtkContainerUnsetFocusChain}
\name{gtkContainerUnsetFocusChain}
\title{gtkContainerUnsetFocusChain}
\description{Removes a focus chain explicitly set with \code{\link{gtkContainerSetFocusChain}}.}
\usage{gtkContainerUnsetFocusChain(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkContainer}}]  a \verb{\link{GtkContainer}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
