\alias{gtkCTreeCollapse}
\name{gtkCTreeCollapse}
\title{gtkCTreeCollapse}
\description{
Collapse one node.
\strong{WARNING: \code{gtk_ctree_collapse} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeCollapse(object, node)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCTree}}] }
\item{\verb{node}}{[\verb{\link{GtkCTreeNode}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
