\alias{gtkCTreeCollapseToDepth}
\name{gtkCTreeCollapseToDepth}
\title{gtkCTreeCollapseToDepth}
\description{
Collapse a node and its children up to the depth given.
\strong{WARNING: \code{gtk_ctree_collapse_to_depth} is deprecated and should not be used in newly-written code.  
Collapse a node and its children up to the depth given.}
}
\usage{gtkCTreeCollapseToDepth(object, node, depth)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
\item{\code{depth}}{[integer] The (absolute) depth up to which to collapse nodes.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
