\alias{gtkCTreeExpand}
\name{gtkCTreeExpand}
\title{gtkCTreeExpand}
\description{
Expand one node.
\strong{WARNING: \code{gtk_ctree_expand} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeExpand(object, node)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCTree}}] }
\item{\verb{node}}{[\verb{\link{GtkCTreeNode}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
