\alias{gtkCTreeExpandToDepth}
\name{gtkCTreeExpandToDepth}
\title{gtkCTreeExpandToDepth}
\description{
Expand a node and its children up to the depth given.
\strong{WARNING: \code{gtk_ctree_expand_to_depth} is deprecated and should not be used in newly-written code.  
Expand a node and its children up to the depth given.}
}
\usage{gtkCTreeExpandToDepth(object, node, depth)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
\item{\code{depth}}{[integer] The (absolute) depth up to which to expand nodes.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
