\alias{gtkCTreeExportToGnode}
\name{gtkCTreeExportToGnode}
\title{gtkCTreeExportToGnode}
\description{
FIXME
\strong{WARNING: \code{gtk_ctree_export_to_gnode} is deprecated and should not be used in newly-written code.  
FIXME}
}
\usage{gtkCTreeExportToGnode(object, parent, sibling, node, func, data = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{parent}}{[GNode] }
\item{\code{sibling}}{[GNode] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
\item{\code{func}}{[\code{\link{GtkCTreeGNodeFunc}}] }
\item{\code{data}}{[R object] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
