\alias{gtkCTreeFindNodePtr}
\name{gtkCTreeFindNodePtr}
\title{gtkCTreeFindNodePtr}
\description{
Finds the node pointer given a \verb{\link{GtkCTreeRow}} structure.
\strong{WARNING: \code{gtk_ctree_find_node_ptr} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeFindNodePtr(object, ctree.row)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCTree}}] }
\item{\verb{ctree.row}}{[\verb{\link{GtkCTreeRow}}] }
}
\value{[\verb{\link{GtkCTreeNode}}] The node pointer.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
