\alias{gtkCTreeInsertGnode}
\name{gtkCTreeInsertGnode}
\title{gtkCTreeInsertGnode}
\description{
FIXME
\strong{WARNING: \code{gtk_ctree_insert_gnode} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeInsertGnode(object, parent, sibling, gnode, func, data = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCTree}}] }
\item{\verb{parent}}{[\verb{\link{GtkCTreeNode}}] }
\item{\verb{sibling}}{[\verb{\link{GtkCTreeNode}}] }
\item{\verb{gnode}}{[GNode] }
\item{\verb{func}}{[\verb{\link{GtkCTreeGNodeFunc}}] }
\item{\verb{data}}{[R object] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
