\alias{gtkCTreeIsAncestor}
\name{gtkCTreeIsAncestor}
\title{gtkCTreeIsAncestor}
\description{

\strong{WARNING: \code{gtk_ctree_is_ancestor} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeIsAncestor(object, node, child)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCTree}}] }
\item{\verb{node}}{[\verb{\link{GtkCTreeNode}}] }
\item{\verb{child}}{[\verb{\link{GtkCTreeNode}}] }
}
\value{[logical] True is \code{node} is an ancestor of \code{child}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
