\alias{gtkCTreeIsHotSpot}
\name{gtkCTreeIsHotSpot}
\title{gtkCTreeIsHotSpot}
\description{

\strong{WARNING: \code{gtk_ctree_is_hot_spot} is deprecated and should not be used in newly-written code.  }
}
\usage{gtkCTreeIsHotSpot(object, x, y)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{x}}{[integer] }
\item{\code{y}}{[integer] }
}
\value{[logical] True if the given coordinates lie on an expander button.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
