\alias{gtkCTreeNodeNth}
\name{gtkCTreeNodeNth}
\title{gtkCTreeNodeNth}
\description{

\strong{WARNING: \code{gtk_ctree_node_nth} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeNth(object, row)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCTree}}] }
\item{\verb{row}}{[numeric] }
}
\value{[\verb{\link{GtkCTreeNode}}] The node corresponding to the \code{row} th row.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
