\alias{gtkCTreeNodeSetBackground}
\name{gtkCTreeNodeSetBackground}
\title{gtkCTreeNodeSetBackground}
\description{

\strong{WARNING: \code{gtk_ctree_node_set_background} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetBackground(object, node, color)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCTree}}] }
\item{\verb{node}}{[\verb{\link{GtkCTreeNode}}] }
\item{\verb{color}}{[\verb{\link{GdkColor}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
