\alias{gtkCTreeNodeSetCellStyle}
\name{gtkCTreeNodeSetCellStyle}
\title{gtkCTreeNodeSetCellStyle}
\description{
Set the style of an individual cell.
\strong{WARNING: \code{gtk_ctree_node_set_cell_style} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetCellStyle(object, node, column, style)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
\item{\code{column}}{[integer] }
\item{\code{style}}{[\code{\link{GtkStyle}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
