\alias{gtkCTreeNodeSetPixmap}
\name{gtkCTreeNodeSetPixmap}
\title{gtkCTreeNodeSetPixmap}
\description{
FIXME
\strong{WARNING: \code{gtk_ctree_node_set_pixmap} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetPixmap(object, node, column, pixmap, mask = NULL)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{node}}{\emph{undocumented
}}
\item{\verb{column}}{\emph{undocumented
}}
\item{\verb{pixmap}}{\emph{undocumented
}}
\item{\verb{mask}}{. \emph{[  \acronym{allow-none}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
