\alias{gtkCTreeNodeSetPixtext}
\name{gtkCTreeNodeSetPixtext}
\title{gtkCTreeNodeSetPixtext}
\description{
FIXME
\strong{WARNING: \code{gtk_ctree_node_set_pixtext} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetPixtext(object, node, column, text, spacing, pixmap, mask = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCTree}}] }
\item{\verb{node}}{[\verb{\link{GtkCTreeNode}}] }
\item{\verb{column}}{[integer] }
\item{\verb{text}}{[character] }
\item{\verb{spacing}}{[raw] }
\item{\verb{pixmap}}{[\verb{\link{GdkPixmap}}] }
\item{\verb{mask}}{[\verb{\link{GdkBitmap}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
