\alias{gtkCTreeNodeSetRowStyle}
\name{gtkCTreeNodeSetRowStyle}
\title{gtkCTreeNodeSetRowStyle}
\description{
Set the style of a row.
\strong{WARNING: \code{gtk_ctree_node_set_row_style} is deprecated and should not be used in newly-written code.  
Set the style of a row.}
}
\usage{gtkCTreeNodeSetRowStyle(object, node, style)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
\item{\code{style}}{[\code{\link{GtkStyle}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
