\alias{gtkCTreeNodeSetShift}
\name{gtkCTreeNodeSetShift}
\title{gtkCTreeNodeSetShift}
\description{
Shift the given cell the given amounts in pixels.
\strong{WARNING: \code{gtk_ctree_node_set_shift} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetShift(object, node, column, vertical, horizontal)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
\item{\code{column}}{[integer] }
\item{\code{vertical}}{[integer] }
\item{\code{horizontal}}{[integer] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
