\alias{gtkCTreeRemoveNode}
\name{gtkCTreeRemoveNode}
\title{gtkCTreeRemoveNode}
\description{
Remove the node and all nodes underneath it from the tree.
\strong{WARNING: \code{gtk_ctree_remove_node} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeRemoveNode(object, node)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCTree}}] The widget.}
\item{\verb{node}}{[\verb{\link{GtkCTreeNode}}] The node to be removed.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
