\alias{gtkCTreeSetDragCompareFunc}
\name{gtkCTreeSetDragCompareFunc}
\title{gtkCTreeSetDragCompareFunc}
\description{
FIXME
\strong{WARNING: \code{gtk_ctree_set_drag_compare_func} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeSetDragCompareFunc(object, cmp.func)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{cmp.func}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
