\alias{gtkCTreeSetIndent}
\name{gtkCTreeSetIndent}
\title{gtkCTreeSetIndent}
\description{

\strong{WARNING: \code{gtk_ctree_set_indent} is deprecated and should not be used in newly-written code.  }
}
\usage{gtkCTreeSetIndent(object, indent)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{indent}}{[integer] The number of pixels to shift the levels of the tree.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
