\alias{gtkCTreeSetLineStyle}
\name{gtkCTreeSetLineStyle}
\title{gtkCTreeSetLineStyle}
\description{

\strong{WARNING: \code{gtk_ctree_set_line_style} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeSetLineStyle(object, line.style)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCTree}}] }
\item{\verb{line.style}}{[\code{\link{GtkCTreeLineStyle}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
