\alias{gtkCTreeSetShowStub}
\name{gtkCTreeSetShowStub}
\title{gtkCTreeSetShowStub}
\description{

\strong{WARNING: \code{gtk_ctree_set_show_stub} is deprecated and should not be used in newly-written code.  }
}
\usage{gtkCTreeSetShowStub(object, show.stub)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{show.stub}}{[logical] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
