\alias{gtkCTreeToggleExpansion}
\name{gtkCTreeToggleExpansion}
\title{gtkCTreeToggleExpansion}
\description{
Toggle a node, i.e. if it is collapsed, expand it and vice versa.
\strong{WARNING: \code{gtk_ctree_toggle_expansion} is deprecated and should not be used in newly-written code.  
Toggle a node, i.e. if it is collapsed, expand it and vice versa.}
}
\usage{gtkCTreeToggleExpansion(object, node)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
