\alias{gtkCTreeUnselect}
\name{gtkCTreeUnselect}
\title{gtkCTreeUnselect}
\description{
Unselect the given node and emit the appropriate signal.
\strong{WARNING: \code{gtk_ctree_unselect} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeUnselect(object, node)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCTree}}] }
\item{\verb{node}}{[\verb{\link{GtkCTreeNode}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
