\alias{gtkCurveGetVector}
\name{gtkCurveGetVector}
\title{gtkCurveGetVector}
\description{Returns a vector of points representing the curve.}
\usage{gtkCurveGetVector(object, veclen)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCurve}}] a \code{\link{GtkCurve}}.}
\item{\code{veclen}}{[integer] the number of points to calculate.}
}
\value{
A list containing the following elements:
\item{\code{vector}}{[numeric] returns the points.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
