\alias{gtkCurveNew}
\name{gtkCurveNew}
\title{gtkCurveNew}
\description{Creates a new \code{\link{GtkCurve}}.}
\usage{gtkCurveNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkCurve}}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
