\alias{gtkCurveSetCurveType}
\name{gtkCurveSetCurveType}
\title{gtkCurveSetCurveType}
\description{Sets the type of the curve. The curve will remain unchanged except when
changing from a free curve to a linear or spline curve, in which case the
curve will be changed as little as possible.}
\usage{gtkCurveSetCurveType(object, type)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCurve}}] a \verb{\link{GtkCurve}}.}
\item{\verb{type}}{[\code{\link{GtkCurveType}}] the type of the curve.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
