\alias{gtkCurveSetRange}
\name{gtkCurveSetRange}
\title{gtkCurveSetRange}
\description{Sets the minimum and maximum x and y values of the curve.
The curve is also reset with a call to \code{\link{gtkCurveReset}}.}
\usage{gtkCurveSetRange(object, min.x, max.x, min.y, max.y)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCurve}}] a \code{\link{GtkCurve}}.}
\item{\code{min.x}}{[numeric] the minimum x value.}
\item{\code{max.x}}{[numeric] the maximum x value.}
\item{\code{min.y}}{[numeric] the minimum y value.}
\item{\code{max.y}}{[numeric] the maximum y value.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
