\alias{gtkCurveSetVector}
\name{gtkCurveSetVector}
\title{gtkCurveSetVector}
\description{Sets the vector of points on the curve.
The curve type is set to \code{GTK_CURVE_TYPE_FREE}.}
\usage{gtkCurveSetVector(object, vector)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCurve}}] a \verb{\link{GtkCurve}}.}
\item{\verb{vector}}{[numeric] the points on the curve.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
