\alias{gtkDialogGetActionArea}
\name{gtkDialogGetActionArea}
\title{gtkDialogGetActionArea}
\description{Returns the action area of \code{dialog}.}
\usage{gtkDialogGetActionArea(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkDialog}}}}
\details{Since 2.14}
\value{[\code{\link{GtkWidget}}] the action area. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
