\alias{gtkDialogGetContentArea}
\name{gtkDialogGetContentArea}
\title{gtkDialogGetContentArea}
\description{Returns the content area of \code{dialog}.}
\usage{gtkDialogGetContentArea(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkDialog}}}}
\details{Since 2.14}
\value{[\code{\link{GtkWidget}}] the content area \code{\link{GtkVBox}}. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
