\alias{gtkDialogGetHasSeparator}
\name{gtkDialogGetHasSeparator}
\title{gtkDialogGetHasSeparator}
\description{Accessor for whether the dialog has a separator.}
\usage{gtkDialogGetHasSeparator(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkDialog}}]  a \code{\link{GtkDialog}}}}
\value{[logical] \code{TRUE} if the dialog has a separator}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
