\alias{gtkDialogNew}
\name{gtkDialogNew}
\title{gtkDialogNew}
\description{Creates a new dialog box. Widgets should not be packed into this \verb{\link{GtkWindow}}
directly, but into the \code{vbox} and \code{action.area}, as described above. }
\usage{gtkDialogNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}] a new \verb{\link{GtkDialog}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
