\alias{gtkDragDestGetTargetList}
\name{gtkDragDestGetTargetList}
\title{gtkDragDestGetTargetList}
\description{Returns the list of targets this widget can accept from
drag-and-drop.}
\usage{gtkDragDestGetTargetList(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}}}
\value{[\verb{\link{GtkTargetList}}]  the \verb{\link{GtkTargetList}}, or \code{NULL} if none}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
