\alias{gtkDragDestGetTrackMotion}
\name{gtkDragDestGetTrackMotion}
\title{gtkDragDestGetTrackMotion}
\description{Returns whether the widget has been configured to always
emit ::drag-motion signals.}
\usage{gtkDragDestGetTrackMotion(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}} that's a drag destination}}
\details{Since 2.10}
\value{[logical] \code{TRUE} if the widget always emits ::drag-motion events}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
