\alias{gtkDragDestUnset}
\name{gtkDragDestUnset}
\title{gtkDragDestUnset}
\description{Clears information about a drop destination set with
\code{\link{gtkDragDestSet}}. The widget will no longer receive
notification of drags.}
\usage{gtkDragDestUnset(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
