\alias{gtkDragSetIconDefault}
\name{gtkDragSetIconDefault}
\title{gtkDragSetIconDefault}
\description{Sets the icon for a particular drag to the default
icon.}
\usage{gtkDragSetIconDefault(object)}
\arguments{\item{\verb{object}}{the context for a drag. (This must be called
with a  context for the source side of a drag)}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
