\alias{gtkDragSourceSetIconStock}
\name{gtkDragSourceSetIconStock}
\title{gtkDragSourceSetIconStock}
\description{Sets the icon that will be used for drags from a particular source
to a stock icon.}
\usage{gtkDragSourceSetIconStock(object, stock.id)}
\arguments{
\item{\code{object}}{[\code{\link{GtkWidget}}]  a \code{\link{GtkWidget}}}
\item{\code{stock.id}}{[character]  the ID of the stock icon to use}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
