\alias{gtkDragSourceUnset}
\name{gtkDragSourceUnset}
\title{gtkDragSourceUnset}
\description{Undoes the effects of \code{\link{gtkDragSourceSet}}.}
\usage{gtkDragSourceUnset(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkWidget}}] a \code{\link{GtkWidget}}}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
