\alias{gtkDragUnhighlight}
\name{gtkDragUnhighlight}
\title{gtkDragUnhighlight}
\description{Removes a highlight set by \code{\link{gtkDragHighlight}} from
a widget.}
\usage{gtkDragUnhighlight(object)}
\arguments{\item{\verb{object}}{a widget to remove the highlight from.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
