\alias{gtkDrawLayout}
\name{gtkDrawLayout}
\title{gtkDrawLayout}
\description{

\strong{WARNING: \code{gtk_draw_layout} is deprecated and should not be used in newly-written code.}
}
\usage{gtkDrawLayout(object, window, state.type, use.text, x, y, layout)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkStyle}}] }
\item{\verb{window}}{[\verb{\link{GdkWindow}}] }
\item{\verb{state.type}}{[\code{\link{GtkStateType}}] }
\item{\verb{use.text}}{[logical] }
\item{\verb{x}}{[integer] }
\item{\verb{y}}{[integer] }
\item{\verb{layout}}{[\verb{\link{PangoLayout}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
