\alias{gtkDrawSlider}
\name{gtkDrawSlider}
\title{gtkDrawSlider}
\description{

\strong{WARNING: \code{gtk_draw_slider} is deprecated and should not be used in newly-written code.}
}
\usage{gtkDrawSlider(object, window, state.type, shadow.type, x, y, width, height, orientation)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkStyle}}] }
\item{\verb{window}}{[\verb{\link{GdkWindow}}] }
\item{\verb{state.type}}{[\code{\link{GtkStateType}}] }
\item{\verb{shadow.type}}{[\code{\link{GtkShadowType}}] }
\item{\verb{x}}{[integer] }
\item{\verb{y}}{[integer] }
\item{\verb{width}}{[integer] }
\item{\verb{height}}{[integer] }
\item{\verb{orientation}}{[\code{\link{GtkOrientation}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
