\alias{gtkEditableCopyClipboard}
\name{gtkEditableCopyClipboard}
\title{gtkEditableCopyClipboard}
\description{Causes the characters in the current selection to
be copied to the clipboard.}
\usage{gtkEditableCopyClipboard(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkEditable}}] a \code{\link{GtkEditable}} widget.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
