\alias{gtkEditableCutClipboard}
\name{gtkEditableCutClipboard}
\title{gtkEditableCutClipboard}
\description{Causes the characters in the current selection to
be copied to the clipboard and then deleted from
the widget.}
\usage{gtkEditableCutClipboard(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkEditable}}] a \code{\link{GtkEditable}} widget.}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
