\alias{gtkEditableGetEditable}
\name{gtkEditableGetEditable}
\title{gtkEditableGetEditable}
\description{Retrieves whether \code{editable} is editable. See
\code{\link{gtkEditableSetEditable}}.}
\usage{gtkEditableGetEditable(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkEditable}}]  a \verb{\link{GtkEditable}}}}
\value{[logical] \code{TRUE} if \code{editable} is editable.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
