\alias{gtkEditablePasteClipboard}
\name{gtkEditablePasteClipboard}
\title{gtkEditablePasteClipboard}
\description{Causes the contents of the clipboard to be pasted into
the given widget at the current cursor position.}
\usage{gtkEditablePasteClipboard(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkEditable}}] a \verb{\link{GtkEditable}} widget.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
