\alias{gtkEntryBufferEmitDeletedText}
\name{gtkEntryBufferEmitDeletedText}
\title{gtkEntryBufferEmitDeletedText}
\description{Used when subclassing \code{\link{GtkEntryBuffer}}}
\usage{gtkEntryBufferEmitDeletedText(object, position, n.chars)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkEntryBuffer}}}
\item{\verb{position}}{position at which text was deleted}
\item{\verb{n.chars}}{number of characters deleted}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
