\alias{gtkEntryBufferEmitInsertedText}
\name{gtkEntryBufferEmitInsertedText}
\title{gtkEntryBufferEmitInsertedText}
\description{Used when subclassing \code{\link{GtkEntryBuffer}}}
\usage{gtkEntryBufferEmitInsertedText(object, position, chars, n.chars)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkEntryBuffer}}}
\item{\verb{position}}{position at which text was inserted}
\item{\verb{chars}}{text that was inserted}
\item{\verb{n.chars}}{number of characters inserted}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
