\alias{gtkEntryBufferGetBytes}
\name{gtkEntryBufferGetBytes}
\title{gtkEntryBufferGetBytes}
\description{Retrieves the length in bytes of the buffer.
See \code{\link{gtkEntryBufferGetLength}}.}
\usage{gtkEntryBufferGetBytes(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEntryBuffer}}}}
\details{Since 2.18}
\value{[numeric] The byte length of the buffer.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
