\alias{gtkEntryBufferGetLength}
\name{gtkEntryBufferGetLength}
\title{gtkEntryBufferGetLength}
\description{Retrieves the length in characters of the buffer.}
\usage{gtkEntryBufferGetLength(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEntryBuffer}}}}
\details{Since 2.18}
\value{[numeric] The number of characters in the buffer.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
