\alias{gtkEntryBufferGetMaxLength}
\name{gtkEntryBufferGetMaxLength}
\title{gtkEntryBufferGetMaxLength}
\description{Retrieves the maximum allowed length of the text in
\code{buffer}. See \code{\link{gtkEntryBufferSetMaxLength}}.}
\usage{gtkEntryBufferGetMaxLength(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEntryBuffer}}}}
\details{Since 2.18}
\value{[integer] the maximum allowed number of characters
in \code{\link{GtkEntryBuffer}}, or 0 if there is no maximum.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
