\alias{gtkEntryCompletionComplete}
\name{gtkEntryCompletionComplete}
\title{gtkEntryCompletionComplete}
\description{Requests a completion operation, or in other words a refiltering of the
current list with completions, using the current key. The completion list
view will be updated accordingly.}
\usage{gtkEntryCompletionComplete(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkEntryCompletion}}.}}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
