\alias{gtkEntryCompletionGetCompletionPrefix}
\name{gtkEntryCompletionGetCompletionPrefix}
\title{gtkEntryCompletionGetCompletionPrefix}
\description{Get the original text entered by the user that triggered
the completion or \code{NULL} if there's no completion ongoing.}
\usage{gtkEntryCompletionGetCompletionPrefix(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEntryCompletion}}}}
\details{Since 2.12}
\value{[character] the prefix for the current completion}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
