\alias{gtkEntryCompletionGetEntry}
\name{gtkEntryCompletionGetEntry}
\title{gtkEntryCompletionGetEntry}
\description{Gets the entry \code{completion} has been attached to.}
\usage{gtkEntryCompletionGetEntry(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkEntryCompletion}}]  A \verb{\link{GtkEntryCompletion}}.}}
\details{  Since  2.4}
\value{[\verb{\link{GtkWidget}}]  The entry \code{completion} has been attached to.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
