\alias{gtkEntryCompletionGetInlineCompletion}
\name{gtkEntryCompletionGetInlineCompletion}
\title{gtkEntryCompletionGetInlineCompletion}
\description{Returns whether the common prefix of the possible completions should
be automatically inserted in the entry.}
\usage{gtkEntryCompletionGetInlineCompletion(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEntryCompletion}}}}
\details{Since 2.6}
\value{[logical] \code{TRUE} if inline completion is turned on}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
