\alias{gtkEntryCompletionGetInlineSelection}
\name{gtkEntryCompletionGetInlineSelection}
\title{gtkEntryCompletionGetInlineSelection}
\description{Returns \code{TRUE} if inline-selection mode is turned on.}
\usage{gtkEntryCompletionGetInlineSelection(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEntryCompletion}}}}
\details{Since 2.12}
\value{[logical] \code{TRUE} if inline-selection mode is on}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
