\alias{gtkEntryCompletionGetModel}
\name{gtkEntryCompletionGetModel}
\title{gtkEntryCompletionGetModel}
\description{Returns the model the \verb{\link{GtkEntryCompletion}} is using as data source.
Returns \code{NULL} if the model is unset.}
\usage{gtkEntryCompletionGetModel(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkEntryCompletion}}]  A \verb{\link{GtkEntryCompletion}}.}}
\details{  Since  2.4}
\value{[\verb{\link{GtkTreeModel}}]  A \verb{\link{GtkTreeModel}}, or \code{NULL} if none is currently being used.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
