\alias{gtkEntryCompletionGetPopupCompletion}
\name{gtkEntryCompletionGetPopupCompletion}
\title{gtkEntryCompletionGetPopupCompletion}
\description{Returns whether the completions should be presented in a popup window.}
\usage{gtkEntryCompletionGetPopupCompletion(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEntryCompletion}}}}
\details{Since 2.6}
\value{[logical] \code{TRUE} if popup completion is turned on}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
