\alias{gtkEntryCompletionGetPopupSetWidth}
\name{gtkEntryCompletionGetPopupSetWidth}
\title{gtkEntryCompletionGetPopupSetWidth}
\description{Returns whether the  completion popup window will be resized to the 
width of the entry.}
\usage{gtkEntryCompletionGetPopupSetWidth(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEntryCompletion}}}}
\details{Since 2.8}
\value{[logical] \code{TRUE} if the popup window will be resized to the width of
the entry}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
