\alias{gtkEntryCompletionGetPopupSingleMatch}
\name{gtkEntryCompletionGetPopupSingleMatch}
\title{gtkEntryCompletionGetPopupSingleMatch}
\description{Returns whether the completion popup window will appear even if there is
only a single match.}
\usage{gtkEntryCompletionGetPopupSingleMatch(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEntryCompletion}}}}
\details{Since 2.8}
\value{[logical] \code{TRUE} if the popup window will appear regardless of the
number of matches.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
