\alias{gtkEntryCompletionInsertActionMarkup}
\name{gtkEntryCompletionInsertActionMarkup}
\title{gtkEntryCompletionInsertActionMarkup}
\description{Inserts an action in \code{completion}'s action item list at position \code{index.}
with markup \code{markup}.}
\usage{gtkEntryCompletionInsertActionMarkup(object, index, markup)}
\arguments{
\item{\code{object}}{[\code{\link{GtkEntryCompletion}}]  A \code{\link{GtkEntryCompletion}}.}
\item{\code{index}}{[integer]  The index of the item to insert.}
\item{\code{markup}}{[character]  Markup of the item to insert.}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
