\alias{gtkEntryCompletionInsertActionText}
\name{gtkEntryCompletionInsertActionText}
\title{gtkEntryCompletionInsertActionText}
\description{Inserts an action in \code{completion}'s action item list at position \code{index.}
with text \code{text}. If you want the action item to have markup, use
\code{\link{gtkEntryCompletionInsertActionMarkup}}.}
\usage{gtkEntryCompletionInsertActionText(object, index, text)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkEntryCompletion}}.}
\item{\verb{index}}{The index of the item to insert.}
\item{\verb{text}}{Text of the item to insert.}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
