\alias{gtkEntryCompletionNew}
\name{gtkEntryCompletionNew}
\title{gtkEntryCompletionNew}
\description{Creates a new \verb{\link{GtkEntryCompletion}} object.}
\usage{gtkEntryCompletionNew()}
\details{  Since  2.4}
\value{[\verb{\link{GtkEntryCompletion}}]  A newly created \verb{\link{GtkEntryCompletion}} object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
