\alias{gtkEntryCompletionSetInlineSelection}
\name{gtkEntryCompletionSetInlineSelection}
\title{gtkEntryCompletionSetInlineSelection}
\description{Sets whether it is possible to cycle through the possible completions
inside the entry.}
\usage{gtkEntryCompletionSetInlineSelection(object, inline.selection)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkEntryCompletion}}]  a \verb{\link{GtkEntryCompletion}}}
\item{\verb{inline.selection}}{[logical] \code{TRUE} to do inline selection}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
