\alias{gtkEntryCompletionSetPopupCompletion}
\name{gtkEntryCompletionSetPopupCompletion}
\title{gtkEntryCompletionSetPopupCompletion}
\description{Sets whether the completions should be presented in a popup window.}
\usage{gtkEntryCompletionSetPopupCompletion(object, popup.completion)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkEntryCompletion}}]  a \verb{\link{GtkEntryCompletion}}}
\item{\verb{popup.completion}}{[logical] \code{TRUE} to do popup completion}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
