\alias{gtkEntryCompletionSetPopupSetWidth}
\name{gtkEntryCompletionSetPopupSetWidth}
\title{gtkEntryCompletionSetPopupSetWidth}
\description{Sets whether the completion popup window will be resized to be the same
width as the entry.}
\usage{gtkEntryCompletionSetPopupSetWidth(object, popup.set.width)}
\arguments{
\item{\code{object}}{[\code{\link{GtkEntryCompletion}}]  a \code{\link{GtkEntryCompletion}}}
\item{\code{popup.set.width}}{[logical]  \code{TRUE} to make the width of the popup the same as the entry}
}
\details{  Since  2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
