\alias{gtkEntryCompletionSetPopupSingleMatch}
\name{gtkEntryCompletionSetPopupSingleMatch}
\title{gtkEntryCompletionSetPopupSingleMatch}
\description{Sets whether the completion popup window will appear even if there is
only a single match. You may want to set this to \code{FALSE} if you
are using inline
completion.}
\usage{gtkEntryCompletionSetPopupSingleMatch(object, popup.single.match)}
\arguments{
\item{\code{object}}{[\code{\link{GtkEntryCompletion}}]  a \code{\link{GtkEntryCompletion}}}
\item{\code{popup.single.match}}{[logical] \code{TRUE} if the popup should appear even for a single
  match}
}
\details{  Since  2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
