\alias{gtkEntryGetAlignment}
\name{gtkEntryGetAlignment}
\title{gtkEntryGetAlignment}
\description{Gets the value set by \code{\link{gtkEntrySetAlignment}}.}
\usage{gtkEntryGetAlignment(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEntry}}}}
\details{Since 2.4}
\value{[numeric] the alignment}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
