\alias{gtkEntryGetBuffer}
\name{gtkEntryGetBuffer}
\title{gtkEntryGetBuffer}
\description{Get the \code{\link{GtkEntryBuffer}} object which holds the text for
this widget.}
\usage{gtkEntryGetBuffer(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEntry}}}}
\details{Since 2.18}
\value{[\code{\link{GtkEntryBuffer}}] A \code{\link{GtkEntryBuffer}} object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
