\alias{gtkEntryGetCursorHadjustment}
\name{gtkEntryGetCursorHadjustment}
\title{gtkEntryGetCursorHadjustment}
\description{Retrieves the horizontal cursor adjustment for the entry. 
See \code{\link{gtkEntrySetCursorHadjustment}}.}
\usage{gtkEntryGetCursorHadjustment(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkEntry}}]  a \verb{\link{GtkEntry}}}}
\details{  Since  2.12}
\value{[\verb{\link{GtkAdjustment}}]  the horizontal cursor adjustment, or \code{NULL} 
  if none has been set.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
