\alias{gtkEntryGetHasFrame}
\name{gtkEntryGetHasFrame}
\title{gtkEntryGetHasFrame}
\description{Gets the value set by \code{\link{gtkEntrySetHasFrame}}.}
\usage{gtkEntryGetHasFrame(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEntry}}}}
\value{[logical] whether the entry has a beveled frame}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
