\alias{gtkEntryGetIconActivatable}
\name{gtkEntryGetIconActivatable}
\title{gtkEntryGetIconActivatable}
\description{Returns whether the icon is activatable.}
\usage{gtkEntryGetIconActivatable(object, icon.pos)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkEntry}}}
\item{\verb{icon.pos}}{Icon position}
}
\details{Since 2.16}
\value{[logical] \code{TRUE} if the icon is activatable.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
