\alias{gtkEntryGetIconSensitive}
\name{gtkEntryGetIconSensitive}
\title{gtkEntryGetIconSensitive}
\description{Returns whether the icon appears sensitive or insensitive.}
\usage{gtkEntryGetIconSensitive(object, icon.pos)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkEntry}}}
\item{\verb{icon.pos}}{Icon position}
}
\details{Since 2.16}
\value{[logical] \code{TRUE} if the icon is sensitive.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
