\alias{gtkEntryGetIconTooltipMarkup}
\name{gtkEntryGetIconTooltipMarkup}
\title{gtkEntryGetIconTooltipMarkup}
\description{Gets the contents of the tooltip on the icon at the specified 
position in \code{entry}.}
\usage{gtkEntryGetIconTooltipMarkup(object, icon.pos)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkEntry}}}
\item{\verb{icon.pos}}{the icon position}
}
\details{Since 2.16}
\value{[character] the tooltip text, or \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
