\alias{gtkEntryGetInnerBorder}
\name{gtkEntryGetInnerBorder}
\title{gtkEntryGetInnerBorder}
\description{This function returns the entry's \verb{"inner-border"} property. See
\code{\link{gtkEntrySetInnerBorder}} for more information.}
\usage{gtkEntryGetInnerBorder(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkEntry}}]  a \verb{\link{GtkEntry}}}}
\details{  Since  2.10}
\value{[\verb{\link{GtkBorder}}]  the entry's \verb{\link{GtkBorder}}, or \code{NULL} if none was set.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
