\alias{gtkEntryGetOverwriteMode}
\name{gtkEntryGetOverwriteMode}
\title{gtkEntryGetOverwriteMode}
\description{Gets the value set by \code{\link{gtkEntrySetOverwriteMode}}.}
\usage{gtkEntryGetOverwriteMode(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEntry}}}}
\details{Since 2.14}
\value{[logical] whether the text is overwritten when typing.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
