\alias{gtkEntryGetProgressFraction}
\name{gtkEntryGetProgressFraction}
\title{gtkEntryGetProgressFraction}
\description{Returns the current fraction of the task that's been completed.
See \code{\link{gtkEntrySetProgressFraction}}.}
\usage{gtkEntryGetProgressFraction(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEntry}}}}
\details{Since 2.16}
\value{[numeric] a fraction from 0.0 to 1.0}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
