\alias{gtkEntryGetProgressPulseStep}
\name{gtkEntryGetProgressPulseStep}
\title{gtkEntryGetProgressPulseStep}
\description{Retrieves the pulse step set with \code{\link{gtkEntrySetProgressPulseStep}}.}
\usage{gtkEntryGetProgressPulseStep(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEntry}}}}
\details{Since 2.16}
\value{[numeric] a fraction from 0.0 to 1.0}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
