\alias{gtkEntryGetVisibility}
\name{gtkEntryGetVisibility}
\title{gtkEntryGetVisibility}
\description{Retrieves whether the text in \code{entry} is visible. See
\code{\link{gtkEntrySetVisibility}}.}
\usage{gtkEntryGetVisibility(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkEntry}}]  a \code{\link{GtkEntry}}}}
\value{[logical]  \code{TRUE} if the text is currently visible}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
