\alias{gtkEntryGetWidthChars}
\name{gtkEntryGetWidthChars}
\title{gtkEntryGetWidthChars}
\description{Gets the value set by \code{\link{gtkEntrySetWidthChars}}.}
\usage{gtkEntryGetWidthChars(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkEntry}}]  a \code{\link{GtkEntry}}}}
\value{[integer]  number of chars to request space for, or negative if unset}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
