\alias{gtkEntryNewWithBuffer}
\name{gtkEntryNewWithBuffer}
\title{gtkEntryNewWithBuffer}
\description{Creates a new entry with the specified text buffer.}
\usage{gtkEntryNewWithBuffer(buffer, show = TRUE)}
\arguments{\item{\verb{buffer}}{The buffer to use for the new \code{\link{GtkEntry}}.}}
\details{Since 2.18}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkEntry}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
