\alias{gtkEntryPrependText}
\name{gtkEntryPrependText}
\title{gtkEntryPrependText}
\description{
Prepends the given text to the contents of th ewidget.
\strong{WARNING: \code{gtk_entry_prepend_text} is deprecated and should not be used in newly-written code.}
}
\usage{gtkEntryPrependText(object, text)}
\arguments{
\item{\code{object}}{[\code{\link{GtkEntry}}] a \code{\link{GtkEntry}}.}
\item{\code{text}}{[character] the text to prepend.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
