\alias{gtkEntrySetAlignment}
\name{gtkEntrySetAlignment}
\title{gtkEntrySetAlignment}
\description{Sets the alignment for the contents of the entry. This controls
the horizontal positioning of the contents when the displayed
text is shorter than the width of the entry.}
\usage{gtkEntrySetAlignment(object, xalign)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkEntry}}}
\item{\verb{xalign}}{The horizontal alignment, from 0 (left) to 1 (right).
Reversed for RTL layouts}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
