\alias{gtkEntrySetBuffer}
\name{gtkEntrySetBuffer}
\title{gtkEntrySetBuffer}
\description{Set the \code{\link{GtkEntryBuffer}} object which holds the text for
this widget.}
\usage{gtkEntrySetBuffer(object, buffer)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkEntry}}}
\item{\verb{buffer}}{a \code{\link{GtkEntryBuffer}}}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
