\alias{gtkEntrySetHasFrame}
\name{gtkEntrySetHasFrame}
\title{gtkEntrySetHasFrame}
\description{Sets whether the entry has a beveled frame around it.}
\usage{gtkEntrySetHasFrame(object, setting)}
\arguments{
\item{\code{object}}{[\code{\link{GtkEntry}}]  a \code{\link{GtkEntry}}}
\item{\code{setting}}{[logical]  new value}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
