\alias{gtkEntrySetIconActivatable}
\name{gtkEntrySetIconActivatable}
\title{gtkEntrySetIconActivatable}
\description{Sets whether the icon is activatable.}
\usage{gtkEntrySetIconActivatable(object, icon.pos, activatable)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkEntry}}}
\item{\verb{icon.pos}}{Icon position}
\item{\verb{activatable}}{\code{TRUE} if the icon should be activatable}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
