\alias{gtkEntrySetIconSensitive}
\name{gtkEntrySetIconSensitive}
\title{gtkEntrySetIconSensitive}
\description{Sets the sensitivity for the specified icon.}
\usage{gtkEntrySetIconSensitive(object, icon.pos, sensitive)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkEntry}}}
\item{\verb{icon.pos}}{Icon position}
\item{\verb{sensitive}}{Specifies whether the icon should appear
sensitive or insensitive}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
