\alias{gtkEntrySetOverwriteMode}
\name{gtkEntrySetOverwriteMode}
\title{gtkEntrySetOverwriteMode}
\description{Sets whether the text is overwritten when typing in the \code{\link{GtkEntry}}.}
\usage{gtkEntrySetOverwriteMode(object, overwrite)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkEntry}}}
\item{\verb{overwrite}}{new value}
}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
