\alias{gtkEntrySetProgressFraction}
\name{gtkEntrySetProgressFraction}
\title{gtkEntrySetProgressFraction}
\description{Causes the entry's progress indicator to "fill in" the given
fraction of the bar. The fraction should be between 0.0 and 1.0,
inclusive.}
\usage{gtkEntrySetProgressFraction(object, fraction)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkEntry}}}
\item{\verb{fraction}}{fraction of the task that's been completed}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
