\alias{gtkEntrySetProgressPulseStep}
\name{gtkEntrySetProgressPulseStep}
\title{gtkEntrySetProgressPulseStep}
\description{Sets the fraction of total entry width to move the progress
bouncing block for each call to \code{\link{gtkEntryProgressPulse}}.}
\usage{gtkEntrySetProgressPulseStep(object, fraction)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkEntry}}}
\item{\verb{fraction}}{fraction between 0.0 and 1.0}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
