\alias{gtkEntrySetText}
\name{gtkEntrySetText}
\title{gtkEntrySetText}
\description{Sets the text in the widget to the given
value, replacing the current contents.}
\usage{gtkEntrySetText(object, text)}
\arguments{
\item{\code{object}}{[\code{\link{GtkEntry}}]  a \code{\link{GtkEntry}}}
\item{\code{text}}{[character]  the new text}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
