\alias{gtkEntryUnsetInvisibleChar}
\name{gtkEntryUnsetInvisibleChar}
\title{gtkEntryUnsetInvisibleChar}
\description{Unsets the invisible char previously set with
\code{\link{gtkEntrySetInvisibleChar}}. So that the
default invisible char is used again.}
\usage{gtkEntryUnsetInvisibleChar(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEntry}}}}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
