\alias{gtkEventBoxGetAboveChild}
\name{gtkEventBoxGetAboveChild}
\title{gtkEventBoxGetAboveChild}
\description{Returns whether the event box window is above or below the
windows of its child. See \code{\link{gtkEventBoxSetAboveChild}} for
details.}
\usage{gtkEventBoxGetAboveChild(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEventBox}}}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if the event box window is above the window
of its child.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
