\alias{gtkEventBoxGetVisibleWindow}
\name{gtkEventBoxGetVisibleWindow}
\title{gtkEventBoxGetVisibleWindow}
\description{Returns whether the event box has a visible window.
See \code{\link{gtkEventBoxSetVisibleWindow}} for details.}
\usage{gtkEventBoxGetVisibleWindow(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEventBox}}}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if the event box window is visible}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
