\alias{gtkEventBoxNew}
\name{gtkEventBoxNew}
\title{gtkEventBoxNew}
\description{Creates a new \code{\link{GtkEventBox}}.}
\usage{gtkEventBoxNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkEventBox}}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
