\alias{gtkExpanderGetExpanded}
\name{gtkExpanderGetExpanded}
\title{gtkExpanderGetExpanded}
\description{Queries a \verb{\link{GtkExpander}} and returns its current state. Returns \code{TRUE}
if the child widget is revealed.}
\usage{gtkExpanderGetExpanded(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkExpander}}] a \verb{\link{GtkExpander}}}}
\details{See \code{\link{gtkExpanderSetExpanded}}.  

  Since  2.4}
\value{[logical]  the current state of the expander.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
