\alias{gtkExpanderGetLabelWidget}
\name{gtkExpanderGetLabelWidget}
\title{gtkExpanderGetLabelWidget}
\description{Retrieves the label widget for the frame. See
\code{\link{gtkExpanderSetLabelWidget}}.}
\usage{gtkExpanderGetLabelWidget(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkExpander}}}}
\details{Since 2.4}
\value{[\code{\link{GtkWidget}}] the label widget, or \code{NULL} if there is none.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
