\alias{gtkExpanderGetUseMarkup}
\name{gtkExpanderGetUseMarkup}
\title{gtkExpanderGetUseMarkup}
\description{Returns whether the label's text is interpreted as marked up with
the Pango text markup
language. See \code{\link{gtkExpanderSetUseMarkup}}.}
\usage{gtkExpanderGetUseMarkup(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkExpander}}}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if the label's text will be parsed for markup}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
