\alias{gtkExpanderGetUseUnderline}
\name{gtkExpanderGetUseUnderline}
\title{gtkExpanderGetUseUnderline}
\description{Returns whether an embedded underline in the expander label indicates a
mnemonic. See \code{\link{gtkExpanderSetUseUnderline}}.}
\usage{gtkExpanderGetUseUnderline(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkExpander}}}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if an embedded underline in the expander label
indicates the mnemonic accelerator keys.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
