\alias{gtkExpanderSetExpanded}
\name{gtkExpanderSetExpanded}
\title{gtkExpanderSetExpanded}
\description{Sets the state of the expander. Set to \code{TRUE}, if you want
the child widget to be revealed, and \code{FALSE} if you want the
child widget to be hidden.}
\usage{gtkExpanderSetExpanded(object, expanded)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkExpander}}]  a \verb{\link{GtkExpander}}}
\item{\verb{expanded}}{[logical]  whether the child widget is revealed}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
