\alias{gtkExpanderSetLabelWidget}
\name{gtkExpanderSetLabelWidget}
\title{gtkExpanderSetLabelWidget}
\description{Set the label widget for the expander. This is the widget
that will appear embedded alongside the expander arrow.}
\usage{gtkExpanderSetLabelWidget(object, label.widget = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkExpander}}]  a \verb{\link{GtkExpander}}}
\item{\verb{label.widget}}{[\verb{\link{GtkWidget}}]  the new label widget}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
