\alias{gtkExpanderSetUseMarkup}
\name{gtkExpanderSetUseMarkup}
\title{gtkExpanderSetUseMarkup}
\description{Sets whether the text of the label contains markup in Pango's text markup
language. See \code{\link{gtkLabelSetMarkup}}.}
\usage{gtkExpanderSetUseMarkup(object, use.markup)}
\arguments{
\item{\code{object}}{[\code{\link{GtkExpander}}]  a \code{\link{GtkExpander}}}
\item{\code{use.markup}}{[logical]  \code{TRUE} if the label's text should be parsed for markup}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
