\alias{gtkFalse}
\name{gtkFalse}
\title{gtkFalse}
\description{Analogical to \code{\link{gtkTrue}} this function does nothing 
but always returns \code{FALSE}.}
\usage{gtkFalse()}
\value{[logical] \code{FALSE}}
\note{
This function is deprecated. You should use \code{GDK_PRIORITY_REDRAW} instead.
This function is deprecated. You should use \code{G_PRIORITY_HIGH} instead.
This function is deprecated. You should use \code{G_PRIORITY_DEFAULT_IDLE} instead.
This function is deprecated. You should use \code{G_PRIORITY_LOW} instead.
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
