\alias{gtkFileChooserButtonGetFocusOnClick}
\name{gtkFileChooserButtonGetFocusOnClick}
\title{gtkFileChooserButtonGetFocusOnClick}
\description{Returns whether the button grabs focus when it is clicked with the mouse.
See \code{\link{gtkFileChooserButtonSetFocusOnClick}}.}
\usage{gtkFileChooserButtonGetFocusOnClick(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFileChooserButton}}}}
\details{Since 2.10}
\value{[logical] \code{TRUE} if the button grabs focus when it is clicked with
the mouse.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
