\alias{gtkFileChooserButtonGetTitle}
\name{gtkFileChooserButtonGetTitle}
\title{gtkFileChooserButtonGetTitle}
\description{Retrieves the title of the browse dialog used by \code{button}. The returned value
should not be modified or freed.}
\usage{gtkFileChooserButtonGetTitle(object)}
\arguments{\item{\verb{object}}{the button widget to examine.}}
\details{Since 2.6}
\value{[character] a pointer to the browse dialog's title.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
