\alias{gtkFileChooserButtonGetWidthChars}
\name{gtkFileChooserButtonGetWidthChars}
\title{gtkFileChooserButtonGetWidthChars}
\description{Retrieves the width in characters of the \code{button} widget's entry and/or label.}
\usage{gtkFileChooserButtonGetWidthChars(object)}
\arguments{\item{\verb{object}}{the button widget to examine.}}
\details{Since 2.6}
\value{[integer] an integer width (in characters) that the button will use to size itself.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
