\alias{gtkFileChooserButtonNewWithBackend}
\name{gtkFileChooserButtonNewWithBackend}
\title{gtkFileChooserButtonNewWithBackend}
\description{Creates a new file-selecting button widget using \code{backend}.}
\usage{gtkFileChooserButtonNewWithBackend(title, action, backend, show = TRUE)}
\arguments{
\item{\verb{title}}{[character]  the title of the browse dialog.}
\item{\verb{action}}{[\code{\link{GtkFileChooserAction}}]  the open mode for the widget.}
\item{\verb{backend}}{[character]  the name of the \verb{GtkFileSystem} backend to use.}
}
\details{  Since  2.6}
\value{[\verb{\link{GtkWidget}}]  a new button widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
