\alias{gtkFileChooserButtonSetTitle}
\name{gtkFileChooserButtonSetTitle}
\title{gtkFileChooserButtonSetTitle}
\description{Modifies the \code{title} of the browse dialog used by \code{button}.}
\usage{gtkFileChooserButtonSetTitle(object, title)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkFileChooserButton}}]  the button widget to modify.}
\item{\verb{title}}{[character]  the new browse dialog title.}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
