\alias{gtkFileChooserButtonSetWidthChars}
\name{gtkFileChooserButtonSetWidthChars}
\title{gtkFileChooserButtonSetWidthChars}
\description{Sets the width (in characters) that \code{button} will use to \code{n.chars}.}
\usage{gtkFileChooserButtonSetWidthChars(object, n.chars)}
\arguments{
\item{\code{object}}{[\code{\link{GtkFileChooserButton}}]  the button widget to examine.}
\item{\code{n.chars}}{[integer]  the new width, in characters.}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
