\alias{gtkFileChooserErrorQuark}
\name{gtkFileChooserErrorQuark}
\title{gtkFileChooserErrorQuark}
\description{Registers an error quark for \code{\link{GtkFileChooser}} if necessary.}
\usage{gtkFileChooserErrorQuark()}
\details{  Since  2.4}
\value{[numeric]  The error quark used for \code{\link{GtkFileChooser}} errors.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
