\alias{gtkFileChooserGetAction}
\name{gtkFileChooserGetAction}
\title{gtkFileChooserGetAction}
\description{Gets the type of operation that the file chooser is performing; see
\code{\link{gtkFileChooserSetAction}}.}
\usage{gtkFileChooserGetAction(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFileChooser}}}}
\details{Since 2.4}
\value{[\code{\link{GtkFileChooserAction}}] the action that the file selector is performing}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
