\alias{gtkFileChooserGetCreateFolders}
\name{gtkFileChooserGetCreateFolders}
\title{gtkFileChooserGetCreateFolders}
\description{Gets whether file choser will offer to create new folders.
See \code{\link{gtkFileChooserSetCreateFolders}}.}
\usage{gtkFileChooserGetCreateFolders(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFileChooser}}}}
\details{Since 2.18}
\value{[logical] \code{TRUE} if the New Folder button should be displayed.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
