\alias{gtkFileChooserGetCurrentFolderFile}
\name{gtkFileChooserGetCurrentFolderFile}
\title{gtkFileChooserGetCurrentFolderFile}
\description{Gets the current folder of \code{chooser} as \code{\link{GFile}}.
See \code{\link{gtkFileChooserGetCurrentFolderUri}}.}
\usage{gtkFileChooserGetCurrentFolderFile(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFileChooser}}}}
\details{Since 2.14}
\value{[\code{\link{GFile}}] the \code{\link{GFile}} for the current folder.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
