\alias{gtkFileChooserGetDoOverwriteConfirmation}
\name{gtkFileChooserGetDoOverwriteConfirmation}
\title{gtkFileChooserGetDoOverwriteConfirmation}
\description{Queries whether a file chooser is set to confirm for overwriting when the user
types a file name that already exists.}
\usage{gtkFileChooserGetDoOverwriteConfirmation(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFileChooser}}}}
\details{Since 2.8}
\value{[logical] \code{TRUE} if the file chooser will present a confirmation dialog;
\code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
