\alias{gtkFileChooserGetExtraWidget}
\name{gtkFileChooserGetExtraWidget}
\title{gtkFileChooserGetExtraWidget}
\description{Gets the current preview widget; see
\code{\link{gtkFileChooserSetExtraWidget}}.}
\usage{gtkFileChooserGetExtraWidget(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFileChooser}}}}
\details{Since 2.4}
\value{[\code{\link{GtkWidget}}] the current extra widget, or \code{NULL}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
