\alias{gtkFileChooserGetLocalOnly}
\name{gtkFileChooserGetLocalOnly}
\title{gtkFileChooserGetLocalOnly}
\description{Gets whether only local files can be selected in the
file selector. See \code{\link{gtkFileChooserSetLocalOnly}}}
\usage{gtkFileChooserGetLocalOnly(object)}
\arguments{\item{\verb{object}}{a \verb{GtkFileChoosre}}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if only local files can be selected.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
