\alias{gtkFileChooserGetPreviewUri}
\name{gtkFileChooserGetPreviewUri}
\title{gtkFileChooserGetPreviewUri}
\description{Gets the URI that should be previewed in a custom preview
widget. See \code{\link{gtkFileChooserSetPreviewWidget}}.}
\usage{gtkFileChooserGetPreviewUri(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFileChooser}}}}
\details{Since 2.4}
\value{[char] the URI for the file to preview, or \code{NULL} if no file is
selected.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
