\alias{gtkFileChooserGetPreviewWidget}
\name{gtkFileChooserGetPreviewWidget}
\title{gtkFileChooserGetPreviewWidget}
\description{Gets the current preview widget; see
\code{\link{gtkFileChooserSetPreviewWidget}}.}
\usage{gtkFileChooserGetPreviewWidget(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkFileChooser}}]  a \verb{\link{GtkFileChooser}}}}
\details{  Since  2.4}
\value{[\verb{\link{GtkWidget}}]  the current preview widget, or \code{NULL}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
