\alias{gtkFileChooserGetPreviewWidgetActive}
\name{gtkFileChooserGetPreviewWidgetActive}
\title{gtkFileChooserGetPreviewWidgetActive}
\description{Gets whether the preview widget set by \code{\link{gtkFileChooserSetPreviewWidget}}
should be shown for the current filename. See
\code{\link{gtkFileChooserSetPreviewWidgetActive}}.}
\usage{gtkFileChooserGetPreviewWidgetActive(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFileChooser}}}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if the preview widget is active for the current filename.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
