\alias{gtkFileChooserGetShowHidden}
\name{gtkFileChooserGetShowHidden}
\title{gtkFileChooserGetShowHidden}
\description{Gets whether hidden files and folders are displayed in the file selector.   
See \code{\link{gtkFileChooserSetShowHidden}}.}
\usage{gtkFileChooserGetShowHidden(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFileChooser}}}}
\details{Since 2.6}
\value{[logical] \code{TRUE} if hidden files and folders are displayed.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
