\alias{gtkFileChooserGetUsePreviewLabel}
\name{gtkFileChooserGetUsePreviewLabel}
\title{gtkFileChooserGetUsePreviewLabel}
\description{Gets whether a stock label should be drawn with the name of the previewed
file.  See \code{\link{gtkFileChooserSetUsePreviewLabel}}.}
\usage{gtkFileChooserGetUsePreviewLabel(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFileChooser}}}}
\value{[logical] \code{TRUE} if the file chooser is set to display a label with the
name of the previewed file, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
