\alias{gtkFileChooserListFilters}
\name{gtkFileChooserListFilters}
\title{gtkFileChooserListFilters}
\description{Lists the current set of user-selectable filters; see
\code{\link{gtkFileChooserAddFilter}}, \code{\link{gtkFileChooserRemoveFilter}}.}
\usage{gtkFileChooserListFilters(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkFileChooser}}]  a \verb{\link{GtkFileChooser}}}}
\details{  Since  2.4}
\value{[list]  a \verb{list} containing the current set of
 user selectable filters.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
