\alias{gtkFileChooserSelectAll}
\name{gtkFileChooserSelectAll}
\title{gtkFileChooserSelectAll}
\description{Selects all the files in the current folder of a file chooser.}
\usage{gtkFileChooserSelectAll(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFileChooser}}}}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
