\alias{gtkFileChooserSetCreateFolders}
\name{gtkFileChooserSetCreateFolders}
\title{gtkFileChooserSetCreateFolders}
\description{Sets whether file choser will offer to create new folders.
This is only relevant if the action is not set to be 
\code{GTK_FILE_CHOOSER_ACTION_OPEN}.}
\usage{gtkFileChooserSetCreateFolders(object, create.folders)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFileChooser}}}
\item{\verb{create.folders}}{\code{TRUE} if the New Folder button should be displayed}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
