\alias{gtkFileChooserUnselectAll}
\name{gtkFileChooserUnselectAll}
\title{gtkFileChooserUnselectAll}
\description{Unselects all the files in the current folder of a file chooser.}
\usage{gtkFileChooserUnselectAll(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFileChooser}}}}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
