\alias{gtkFileChooserUnselectFile}
\name{gtkFileChooserUnselectFile}
\title{gtkFileChooserUnselectFile}
\description{Unselects the file referred to by \code{file}. If the file is not in the current
directory, does not exist, or is otherwise not currently selected, does nothing.}
\usage{gtkFileChooserUnselectFile(object, file)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFileChooser}}}
\item{\verb{file}}{a \code{\link{GFile}}}
}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
