\alias{gtkFileChooserUnselectUri}
\name{gtkFileChooserUnselectUri}
\title{gtkFileChooserUnselectUri}
\description{Unselects the file referred to by \code{uri}. If the file
is not in the current directory, does not exist, or
is otherwise not currently selected, does nothing.}
\usage{gtkFileChooserUnselectUri(object, uri)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFileChooser}}}
\item{\verb{uri}}{the URI to unselect}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
