\alias{gtkFileChooserWidgetNew}
\name{gtkFileChooserWidgetNew}
\title{gtkFileChooserWidgetNew}
\description{Creates a new \verb{\link{GtkFileChooserWidget}}.  This is a file chooser widget that can
be embedded in custom windows, and it is the same widget that is used by
\verb{\link{GtkFileChooserDialog}}.}
\usage{gtkFileChooserWidgetNew(action, show = TRUE)}
\arguments{\item{\verb{action}}{[\code{\link{GtkFileChooserAction}}]  Open or save mode for the widget}}
\details{  Since  2.4}
\value{[\verb{\link{GtkWidget}}]  a new \verb{\link{GtkFileChooserWidget}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
