\alias{gtkFileFilterGetName}
\name{gtkFileFilterGetName}
\title{gtkFileFilterGetName}
\description{Gets the human-readable name for the filter. See \code{\link{gtkFileFilterSetName}}.}
\usage{gtkFileFilterGetName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFileFilter}}}}
\details{Since 2.4}
\value{[character] The human-readable name of the filter,
or \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
