\alias{gtkFileFilterSetName}
\name{gtkFileFilterSetName}
\title{gtkFileFilterSetName}
\description{Sets the human-readable name of the filter; this is the string
that will be displayed in the file selector user interface if
there is a selectable list of filters.}
\usage{gtkFileFilterSetName(object, name)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFileFilter}}}
\item{\verb{name}}{the human-readable-name for the filter, or \code{NULL}
to remove any existing name. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
