\alias{gtkFileSelectionGetSelectMultiple}
\name{gtkFileSelectionGetSelectMultiple}
\title{gtkFileSelectionGetSelectMultiple}
\description{
Determines whether or not the user is allowed to select multiple files in
the file list. See \code{\link{gtkFileSelectionSetSelectMultiple}}.
\strong{WARNING: \code{gtk_file_selection_get_select_multiple} is deprecated and should not be used in newly-written code.}
}
\usage{gtkFileSelectionGetSelectMultiple(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFileSelection}}}}
\value{[logical] \code{TRUE} if the user is allowed to select multiple files in the
file list}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
