\alias{gtkFileSelectionHideFileopButtons}
\name{gtkFileSelectionHideFileopButtons}
\title{gtkFileSelectionHideFileopButtons}
\description{Hides the file operation buttons that normally appear at the top of the dialog. Useful if you wish to create a custom file selector, based on \code{\link{GtkFileSelection}}.}
\usage{gtkFileSelectionHideFileopButtons(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkFileSelection}}] a \code{\link{GtkFileSelection}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
