\alias{gtkFileSelectionShowFileopButtons}
\name{gtkFileSelectionShowFileopButtons}
\title{gtkFileSelectionShowFileopButtons}
\description{Shows the file operation buttons, if they have previously been hidden. The rest of the widgets in the dialog will be resized accordingly.}
\usage{gtkFileSelectionShowFileopButtons(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkFileSelection}}] a \code{\link{GtkFileSelection}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
