\alias{gtkFixedGetHasWindow}
\name{gtkFixedGetHasWindow}
\title{gtkFixedGetHasWindow}
\description{Gets whether the \verb{\link{GtkFixed}} has its own \verb{\link{GdkWindow}}.
See \code{gdkFixedSetHasWindow()}.}
\usage{gtkFixedGetHasWindow(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkFixed}}]  a \verb{\link{GtkWidget}}}}
\value{[logical] \code{TRUE} if \code{fixed} has its own window.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
