\alias{gtkFixedMove}
\name{gtkFixedMove}
\title{gtkFixedMove}
\description{Moves a child of a \code{\link{GtkFixed}} container to the given position.}
\usage{gtkFixedMove(object, widget, x, y)}
\arguments{
\item{\code{object}}{[\code{\link{GtkFixed}}] a \code{\link{GtkFixed}}.}
\item{\code{widget}}{[\code{\link{GtkWidget}}] the child widget.}
\item{\code{x}}{[integer] the horizontal position to move the widget to.}
\item{\code{y}}{[integer] the vertical position to move the widget to.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
