\alias{gtkFixedPut}
\name{gtkFixedPut}
\title{gtkFixedPut}
\description{Adds a widget to a \code{\link{GtkFixed}} container at the given position.}
\usage{gtkFixedPut(object, widget, x, y)}
\arguments{
\item{\code{object}}{[\code{\link{GtkFixed}}] a \code{\link{GtkFixed}}.}
\item{\code{widget}}{[\code{\link{GtkWidget}}] the widget to add.}
\item{\code{x}}{[integer] the horizontal position to place the widget at.}
\item{\code{y}}{[integer] the vertical position to place the widget at.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
