\alias{gtkFontButtonGetShowSize}
\name{gtkFontButtonGetShowSize}
\title{gtkFontButtonGetShowSize}
\description{Returns whether the font size will be shown in the label.}
\usage{gtkFontButtonGetShowSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFontButton}}}}
\details{Since 2.4}
\value{[logical] whether the font size will be shown in the label.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
