\alias{gtkFontButtonGetShowStyle}
\name{gtkFontButtonGetShowStyle}
\title{gtkFontButtonGetShowStyle}
\description{Returns whether the name of the font style will be shown in the label.}
\usage{gtkFontButtonGetShowStyle(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFontButton}}}}
\details{Since 2.4}
\value{[logical] whether the font style will be shown in the label.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
