\alias{gtkFontButtonGetTitle}
\name{gtkFontButtonGetTitle}
\title{gtkFontButtonGetTitle}
\description{Retrieves the title of the font selection dialog.}
\usage{gtkFontButtonGetTitle(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkFontButton}}]  a \verb{\link{GtkFontButton}}}}
\details{  Since  2.4}
\value{[character] .}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
