\alias{gtkFontButtonGetUseFont}
\name{gtkFontButtonGetUseFont}
\title{gtkFontButtonGetUseFont}
\description{Returns whether the selected font is used in the label.}
\usage{gtkFontButtonGetUseFont(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFontButton}}}}
\details{Since 2.4}
\value{[logical] whether the selected font is used in the label.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
