\alias{gtkFontButtonGetUseSize}
\name{gtkFontButtonGetUseSize}
\title{gtkFontButtonGetUseSize}
\description{Returns whether the selected size is used in the label.}
\usage{gtkFontButtonGetUseSize(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkFontButton}}]  a \verb{\link{GtkFontButton}}}}
\details{  Since  2.4}
\value{[logical]  whether the selected size is used in the label.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
