\alias{gtkFontButtonNew}
\name{gtkFontButtonNew}
\title{gtkFontButtonNew}
\description{Creates a new font picker widget.}
\usage{gtkFontButtonNew(show = TRUE)}
\details{Since 2.4}
\value{[\code{\link{GtkWidget}}] a new font picker widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
