\alias{gtkFontButtonNewWithFont}
\name{gtkFontButtonNewWithFont}
\title{gtkFontButtonNewWithFont}
\description{Creates a new font picker widget.}
\usage{gtkFontButtonNewWithFont(fontname)}
\arguments{\item{\verb{fontname}}{Name of font to display in font selection dialog}}
\details{Since 2.4}
\value{[\code{\link{GtkWidget}}] a new font picker widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
