\alias{gtkFontButtonSetFontName}
\name{gtkFontButtonSetFontName}
\title{gtkFontButtonSetFontName}
\description{Sets or updates the currently-displayed font in font picker dialog.}
\usage{gtkFontButtonSetFontName(object, fontname)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFontButton}}}
\item{\verb{fontname}}{Name of font to display in font selection dialog}
}
\details{Since 2.4}
\value{[logical] Return value of \code{\link{gtkFontSelectionDialogSetFontName}} if the
font selection dialog exists, otherwise \code{FALSE}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
