\alias{gtkFontButtonSetShowSize}
\name{gtkFontButtonSetShowSize}
\title{gtkFontButtonSetShowSize}
\description{If \code{show.size} is \code{TRUE}, the font size will be displayed along with the name of the selected font.}
\usage{gtkFontButtonSetShowSize(object, show.size)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkFontButton}}]  a \verb{\link{GtkFontButton}}}
\item{\verb{show.size}}{[logical] \code{TRUE} if font size should be displayed in dialog.}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
