\alias{gtkFontButtonSetShowStyle}
\name{gtkFontButtonSetShowStyle}
\title{gtkFontButtonSetShowStyle}
\description{If \code{show.style} is \code{TRUE}, the font style will be displayed along with name of the selected font.}
\usage{gtkFontButtonSetShowStyle(object, show.style)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFontButton}}}
\item{\verb{show.style}}{\code{TRUE} if font style should be displayed in label.}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
