\alias{gtkFontButtonSetTitle}
\name{gtkFontButtonSetTitle}
\title{gtkFontButtonSetTitle}
\description{Sets the title for the font selection dialog.}
\usage{gtkFontButtonSetTitle(object, title)}
\arguments{
\item{\code{object}}{[\code{\link{GtkFontButton}}]  a \code{\link{GtkFontButton}}}
\item{\code{title}}{[character]  a string containing the font selection dialog title}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
