\alias{gtkFontButtonSetUseSize}
\name{gtkFontButtonSetUseSize}
\title{gtkFontButtonSetUseSize}
\description{If \code{use.size} is \code{TRUE}, the font name will be written using the selected size.}
\usage{gtkFontButtonSetUseSize(object, use.size)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFontButton}}}
\item{\verb{use.size}}{If \code{TRUE}, font name will be written using the selected size.}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
