\alias{gtkFontSelectionDialogGetApplyButton}
\name{gtkFontSelectionDialogGetApplyButton}
\title{gtkFontSelectionDialogGetApplyButton}
\description{
Obtains a button. The button doesn't have any function.
\strong{WARNING: \code{gtk_font_selection_dialog_get_apply_button} has been deprecated since version 2.16 and should not be used in newly-written code. Don't use this function.}
}
\usage{gtkFontSelectionDialogGetApplyButton(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFontSelectionDialog}}}}
\details{Since 2.14}
\value{[\code{\link{GtkWidget}}] a \code{\link{GtkWidget}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
