\alias{gtkFontSelectionDialogGetCancelButton}
\name{gtkFontSelectionDialogGetCancelButton}
\title{gtkFontSelectionDialogGetCancelButton}
\description{Gets the 'Cancel' button.}
\usage{gtkFontSelectionDialogGetCancelButton(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFontSelectionDialog}}}}
\details{Since 2.14}
\value{[\code{\link{GtkWidget}}] the \code{\link{GtkWidget}} used in the dialog for the 'Cancel' button.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
