\alias{gtkFontSelectionDialogGetFontName}
\name{gtkFontSelectionDialogGetFontName}
\title{gtkFontSelectionDialogGetFontName}
\description{Gets the currently-selected font name.}
\usage{gtkFontSelectionDialogGetFontName(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkFontSelectionDialog}}] a \code{\link{GtkFontSelectionDialog}}.}}
\value{[character] the currently-selected font name, or \code{NULL} if no font is selected.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
