\alias{gtkFontSelectionDialogGetOkButton}
\name{gtkFontSelectionDialogGetOkButton}
\title{gtkFontSelectionDialogGetOkButton}
\description{Gets the 'OK' button.}
\usage{gtkFontSelectionDialogGetOkButton(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFontSelectionDialog}}}}
\details{Since 2.14}
\value{[\code{\link{GtkWidget}}] the \code{\link{GtkWidget}} used in the dialog for the 'OK' button.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
