\alias{gtkFontSelectionDialogSetFontName}
\name{gtkFontSelectionDialogSetFontName}
\title{gtkFontSelectionDialogSetFontName}
\description{Sets the currently-selected font.}
\usage{gtkFontSelectionDialogSetFontName(object, fontname)}
\arguments{
\item{\code{object}}{[\code{\link{GtkFontSelectionDialog}}] a \code{\link{GtkFontSelectionDialog}}.}
\item{\code{fontname}}{[character] a fontname.}
}
\value{[logical] \code{TRUE} if the font was found.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
