\alias{gtkFontSelectionDialogSetPreviewText}
\name{gtkFontSelectionDialogSetPreviewText}
\title{gtkFontSelectionDialogSetPreviewText}
\description{Sets the text displayed in the preview area.}
\usage{gtkFontSelectionDialogSetPreviewText(object, text)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkFontSelectionDialog}}] a \verb{\link{GtkFontSelectionDialog}}.}
\item{\verb{text}}{[character] the text to display in the preview area.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
