\alias{gtkFontSelectionGetFace}
\name{gtkFontSelectionGetFace}
\title{gtkFontSelectionGetFace}
\description{Gets the \code{\link{PangoFontFace}} representing the selected font group
details (i.e. family, slant, weight, width, etc).}
\usage{gtkFontSelectionGetFace(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFontSelection}}}}
\details{Since 2.14}
\value{[\code{\link{PangoFontFace}}] A \code{\link{PangoFontFace}} representing the selected font
group details.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
