\alias{gtkFontSelectionGetFaceList}
\name{gtkFontSelectionGetFaceList}
\title{gtkFontSelectionGetFaceList}
\description{This returns the \code{\link{GtkTreeView}} which lists all styles available for
the selected font. For example, 'Regular', 'Bold', etc.}
\usage{gtkFontSelectionGetFaceList(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFontSelection}}}}
\details{Since 2.14}
\value{[\code{\link{GtkWidget}}] A \code{\link{GtkWidget}} that is part of \code{fontsel}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
