\alias{gtkFontSelectionGetFamily}
\name{gtkFontSelectionGetFamily}
\title{gtkFontSelectionGetFamily}
\description{Gets the \code{\link{PangoFontFamily}} representing the selected font family.}
\usage{gtkFontSelectionGetFamily(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFontSelection}}}}
\details{Since 2.14}
\value{[\code{\link{PangoFontFamily}}] A \code{\link{PangoFontFamily}} representing the selected font
family. Font families are a collection of font faces.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
