\alias{gtkFontSelectionGetFamilyList}
\name{gtkFontSelectionGetFamilyList}
\title{gtkFontSelectionGetFamilyList}
\description{This returns the \code{\link{GtkTreeView}} that lists font families, for
example, 'Sans', 'Serif', etc.}
\usage{gtkFontSelectionGetFamilyList(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFontSelection}}}}
\details{Since 2.14}
\value{[\code{\link{GtkWidget}}] A \code{\link{GtkWidget}} that is part of \code{fontsel}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
