\alias{gtkFontSelectionGetFont}
\name{gtkFontSelectionGetFont}
\title{gtkFontSelectionGetFont}
\description{
Gets the currently-selected font.
\strong{WARNING: \code{gtk_font_selection_get_font} is deprecated and should not be used in newly-written code.  
Gets the currently-selected font.}
}
\usage{gtkFontSelectionGetFont(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkFontSelection}}] a \code{\link{GtkFontSelection}}.}}
\value{[\code{\link{GdkFont}}] the currently-selected font, or NULL if no font is selected.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
