\alias{gtkFontSelectionGetPreviewEntry}
\name{gtkFontSelectionGetPreviewEntry}
\title{gtkFontSelectionGetPreviewEntry}
\description{This returns the \code{\link{GtkEntry}} used to display the font as a preview.}
\usage{gtkFontSelectionGetPreviewEntry(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFontSelection}}}}
\details{Since 2.14}
\value{[\code{\link{GtkWidget}}] A \code{\link{GtkWidget}} that is part of \code{fontsel}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
