\alias{gtkFontSelectionGetSize}
\name{gtkFontSelectionGetSize}
\title{gtkFontSelectionGetSize}
\description{The selected font size.}
\usage{gtkFontSelectionGetSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFontSelection}}}}
\details{Since 2.14}
\value{[integer] A n integer representing the selected font size,
or -1 if no font size is selected.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
