\alias{gtkFontSelectionGetSizeEntry}
\name{gtkFontSelectionGetSizeEntry}
\title{gtkFontSelectionGetSizeEntry}
\description{This returns the \code{\link{GtkEntry}} used to allow the user to edit the font
number manually instead of selecting it from the list of font sizes.}
\usage{gtkFontSelectionGetSizeEntry(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFontSelection}}}}
\details{Since 2.14}
\value{[\code{\link{GtkWidget}}] A \code{\link{GtkWidget}} that is part of \code{fontsel}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
