\alias{gtkFontSelectionGetSizeList}
\name{gtkFontSelectionGetSizeList}
\title{gtkFontSelectionGetSizeList}
\description{This returns the \verb{GtkTreeeView} used to list font sizes.}
\usage{gtkFontSelectionGetSizeList(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFontSelection}}}}
\details{Since 2.14}
\value{[\code{\link{GtkWidget}}] A \code{\link{GtkWidget}} that is part of \code{fontsel}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
