\alias{gtkFontSelectionNew}
\name{gtkFontSelectionNew}
\title{gtkFontSelectionNew}
\description{Creates a new \code{\link{GtkFontSelection}}.}
\usage{gtkFontSelectionNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkFontSelection}}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
