\alias{gtkFrameGetLabelAlign}
\name{gtkFrameGetLabelAlign}
\title{gtkFrameGetLabelAlign}
\description{Retrieves the X and Y alignment of the frame's label. See
\code{\link{gtkFrameSetLabelAlign}}.}
\usage{gtkFrameGetLabelAlign(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkFrame}}]  a \verb{\link{GtkFrame}}}}
\value{
A list containing the following elements:
\item{\verb{xalign}}{[numeric]  location to store X alignment of frame's label, or \code{NULL}}
\item{\verb{yalign}}{[numeric]  location to store X alignment of frame's label, or \code{NULL}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
