\alias{gtkFrameGetShadowType}
\name{gtkFrameGetShadowType}
\title{gtkFrameGetShadowType}
\description{Retrieves the shadow type of the frame. See
\code{\link{gtkFrameSetShadowType}}.}
\usage{gtkFrameGetShadowType(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFrame}}}}
\value{[\code{\link{GtkShadowType}}] the current shadow type of the frame.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
