\alias{gtkFrameNew}
\name{gtkFrameNew}
\title{gtkFrameNew}
\description{Creates a new \verb{\link{GtkFrame}}, with optional label \code{label}.
If \code{label} is \code{NULL}, the label is omitted.}
\usage{gtkFrameNew(label = NULL, show = TRUE)}
\arguments{\item{\verb{label}}{[character]  the text to use as the label of the frame}}
\value{[\verb{\link{GtkWidget}}]  a new \verb{\link{GtkFrame}} widget}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
