\alias{gtkFrameSetLabelWidget}
\name{gtkFrameSetLabelWidget}
\title{gtkFrameSetLabelWidget}
\description{Sets the label widget for the frame. This is the widget that
will appear embedded in the top edge of the frame as a
title.}
\usage{gtkFrameSetLabelWidget(object, label.widget)}
\arguments{
\item{\code{object}}{[\code{\link{GtkFrame}}]  a \code{\link{GtkFrame}}}
\item{\code{label.widget}}{[\code{\link{GtkWidget}}]  the new label widget}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
