\alias{gtkFrameSetShadowType}
\name{gtkFrameSetShadowType}
\title{gtkFrameSetShadowType}
\description{Sets the shadow type for \code{frame}.}
\usage{gtkFrameSetShadowType(object, type)}
\arguments{
\item{\code{object}}{[\code{\link{GtkFrame}}]  a \code{\link{GtkFrame}}}
\item{\code{type}}{[\code{\link{GtkShadowType}}]  the new \code{\link{GtkShadowType}}}
}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
