\alias{gtkGcRelease}
\name{gtkGcRelease}
\title{gtkGcRelease}
\description{Releases a \code{\link{GdkGC}} allocated using \code{\link{gtkGcGet}}.}
\usage{gtkGcRelease(gc)}
\arguments{\item{\code{gc}}{[\code{\link{GdkGC}}] a \code{\link{GdkGC}}.}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
