\alias{gtkGetCurrentEventTime}
\name{gtkGetCurrentEventTime}
\title{gtkGetCurrentEventTime}
\description{If there is a current event and it has a timestamp, return that
timestamp, otherwise return \code{GDK_CURRENT_TIME}.}
\usage{gtkGetCurrentEventTime()}
\value{[numeric] the timestamp from the current event, or \code{GDK_CURRENT_TIME}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
