\alias{gtkGrabGetCurrent}
\name{gtkGrabGetCurrent}
\title{gtkGrabGetCurrent}
\description{Queries the current grab of the default window group. }
\usage{gtkGrabGetCurrent()}
\value{[\code{\link{GtkWidget}}] The widget which currently has the grab or \code{NULL} if no grab is active.}
\note{
This function is deprecated. You should use \code{GDK_PRIORITY_REDRAW} instead.
This function is deprecated. You should use \code{G_PRIORITY_HIGH} instead.
This function is deprecated. You should use \code{G_PRIORITY_DEFAULT_IDLE} instead.
This function is deprecated. You should use \code{G_PRIORITY_LOW} instead.
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
