\alias{gtkGrabRemove}
\name{gtkGrabRemove}
\title{gtkGrabRemove}
\description{Removes the grab from the given widget. You have to pair calls to \code{\link{gtkGrabAdd}}
and \code{\link{gtkGrabRemove}}.}
\usage{gtkGrabRemove(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkWidget}}] The widget which gives up the grab.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
