\alias{gtkHandleBoxGetChildDetached}
\name{gtkHandleBoxGetChildDetached}
\title{gtkHandleBoxGetChildDetached}
\description{Whether the handlebox's child is currently detached.}
\usage{gtkHandleBoxGetChildDetached(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkHandleBox}}}}
\details{Since 2.14}
\value{[logical] \code{TRUE} if the child is currently detached, otherwise \code{FALSE}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
