\alias{gtkHandleBoxGetHandlePosition}
\name{gtkHandleBoxGetHandlePosition}
\title{gtkHandleBoxGetHandlePosition}
\description{Gets the handle position of the handle box. See
\code{\link{gtkHandleBoxSetHandlePosition}}.}
\usage{gtkHandleBoxGetHandlePosition(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkHandleBox}}}}
\value{[\code{\link{GtkPositionType}}] the current handle position.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
