\alias{gtkHandleBoxGetShadowType}
\name{gtkHandleBoxGetShadowType}
\title{gtkHandleBoxGetShadowType}
\description{Gets the type of shadow drawn around the handle box. See
\code{\link{gtkHandleBoxSetShadowType}}.}
\usage{gtkHandleBoxGetShadowType(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkHandleBox}}]  a \verb{\link{GtkHandleBox}}}}
\value{[\code{\link{GtkShadowType}}]  the type of shadow currently drawn around the handle box.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
