\alias{gtkHandleBoxGetSnapEdge}
\name{gtkHandleBoxGetSnapEdge}
\title{gtkHandleBoxGetSnapEdge}
\description{Gets the edge used for determining reattachment of the handle box. See
\code{\link{gtkHandleBoxSetSnapEdge}}.}
\usage{gtkHandleBoxGetSnapEdge(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkHandleBox}}}}
\value{[\code{\link{GtkPositionType}}] the edge used for determining reattachment, or (GtkPositionType)-1 if this
is determined (as per default) from the handle position.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
