\alias{gtkHandleBoxSetHandlePosition}
\name{gtkHandleBoxSetHandlePosition}
\title{gtkHandleBoxSetHandlePosition}
\description{Sets the side of the handlebox where the handle is drawn.}
\usage{gtkHandleBoxSetHandlePosition(object, position)}
\arguments{
\item{\code{object}}{[\code{\link{GtkHandleBox}}] a \code{\link{GtkHandleBox}}}
\item{\code{position}}{[\code{\link{GtkPositionType}}] the side of the handlebox where the handle should be drawn.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
