\alias{gtkHandleBoxSetShadowType}
\name{gtkHandleBoxSetShadowType}
\title{gtkHandleBoxSetShadowType}
\description{Sets the type of shadow to be drawn around the border 
of the handle box.}
\usage{gtkHandleBoxSetShadowType(object, type)}
\arguments{
\item{\code{object}}{[\code{\link{GtkHandleBox}}] a \code{\link{GtkHandleBox}}}
\item{\code{type}}{[\code{\link{GtkShadowType}}] the shadow type.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
