\alias{gtkHBoxNew}
\name{gtkHBoxNew}
\title{gtkHBoxNew}
\description{Creates a new GtkHBox.}
\usage{gtkHBoxNew(homogeneous = NULL, spacing = NULL, show = TRUE)}
\arguments{
\item{\code{homogeneous}}{[logical] \code{TRUE} if all children are to be given equal space allotments.}
\item{\code{spacing}}{[integer] the number of pixels to place by default between children.}
}
\value{[\code{\link{GtkWidget}}] a new GtkHBox.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
