\alias{gtkHButtonBoxGetSpacingDefault}
\name{gtkHButtonBoxGetSpacingDefault}
\title{gtkHButtonBoxGetSpacingDefault}
\description{
Retrieves the current default spacing for horizontal button boxes. This is the number of pixels
to be placed between the buttons when they are arranged.
\strong{WARNING: \code{gtk_hbutton_box_get_spacing_default} is deprecated and should not be used in newly-written code.}
}
\usage{gtkHButtonBoxGetSpacingDefault()}
\value{[integer] the default number of pixels between buttons.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
