\alias{gtkHButtonBoxNew}
\name{gtkHButtonBoxNew}
\title{gtkHButtonBoxNew}
\description{Creates a new horizontal button box.}
\usage{gtkHButtonBoxNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}] a new button box \verb{\link{GtkWidget}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
