\alias{gtkHButtonBoxSetLayoutDefault}
\name{gtkHButtonBoxSetLayoutDefault}
\title{gtkHButtonBoxSetLayoutDefault}
\description{
Sets a new layout mode that will be used by all button boxes.
\strong{WARNING: \code{gtk_hbutton_box_set_layout_default} is deprecated and should not be used in newly-written code.}
}
\usage{gtkHButtonBoxSetLayoutDefault(layout)}
\arguments{\item{\verb{layout}}{a new \code{\link{GtkButtonBoxStyle}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
