\alias{gtkHPanedNew}
\name{gtkHPanedNew}
\title{gtkHPanedNew}
\description{Create a new \code{\link{GtkHPaned}}}
\usage{gtkHPanedNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] the new \code{\link{GtkHPaned}}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
