\alias{gtkHRulerNew}
\name{gtkHRulerNew}
\title{gtkHRulerNew}
\description{Creates a new horizontal ruler.}
\usage{gtkHRulerNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkHRuler}}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
