\alias{gtkHScaleNew}
\name{gtkHScaleNew}
\title{gtkHScaleNew}
\description{Creates a new \code{\link{GtkHScale}}.}
\usage{gtkHScaleNew(adjustment = NULL, show = TRUE)}
\arguments{\item{\verb{adjustment}}{the \code{\link{GtkAdjustment}} which sets the range of the scale.}}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkHScale}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
