\alias{gtkHScrollbarNew}
\name{gtkHScrollbarNew}
\title{gtkHScrollbarNew}
\description{Creates a new horizontal scrollbar.}
\usage{gtkHScrollbarNew(adjustment = NULL, show = TRUE)}
\arguments{\item{\code{adjustment}}{[\code{\link{GtkAdjustment}}] the \code{\link{GtkAdjustment}} to use, or \code{NULL} to create a new adjustment.}}
\value{[\code{\link{GtkWidget}}] the new \code{\link{GtkHScrollbar}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
