\alias{gtkHSeparatorNew}
\name{gtkHSeparatorNew}
\title{gtkHSeparatorNew}
\description{Creates a new \code{\link{GtkHSeparator}}.}
\usage{gtkHSeparatorNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkHSeparator}}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
