\alias{gtkHSVGetColor}
\name{gtkHSVGetColor}
\title{gtkHSVGetColor}
\description{Queries the current color in an HSV color selector.
Returned values will be in the [0.0, 1.0] range.}
\usage{gtkHSVGetColor(object, h, s, v)}
\arguments{
\item{\verb{object}}{An HSV color selector}
\item{\verb{h}}{Return value for the hue}
\item{\verb{s}}{Return value for the saturation}
\item{\verb{v}}{Return value for the value}
}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
