\alias{gtkHSVGetMetrics}
\name{gtkHSVGetMetrics}
\title{gtkHSVGetMetrics}
\description{Queries the size and ring width of an HSV color selector.}
\usage{gtkHSVGetMetrics(object, size, ring.width)}
\arguments{
\item{\verb{object}}{An HSV color selector}
\item{\verb{size}}{Return value for the diameter of the hue ring}
\item{\verb{ring.width}}{Return value for the width of the hue ring}
}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
