\alias{gtkHSVSetColor}
\name{gtkHSVSetColor}
\title{gtkHSVSetColor}
\description{Sets the current color in an HSV color selector.
Color component values must be in the [0.0, 1.0] range.}
\usage{gtkHSVSetColor(object, h, s, v)}
\arguments{
\item{\verb{object}}{An HSV color selector}
\item{\verb{h}}{Hue}
\item{\verb{s}}{Saturation}
\item{\verb{v}}{Value}
}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
