\alias{gtkHSVSetMetrics}
\name{gtkHSVSetMetrics}
\title{gtkHSVSetMetrics}
\description{Sets the size and ring width of an HSV color selector.}
\usage{gtkHSVSetMetrics(object, size, ring.width)}
\arguments{
\item{\verb{object}}{An HSV color selector}
\item{\verb{size}}{Diameter for the hue ring}
\item{\verb{ring.width}}{Width of the hue ring}
}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
