\alias{gtkHSVToRgb}
\name{gtkHSVToRgb}
\title{gtkHSVToRgb}
\description{Converts a color from HSV space to RGB.
Input values must be in the [0.0, 1.0] range; 
output values will be in the same range.}
\usage{gtkHSVToRgb(h, s, r, g, b)}
\arguments{
\item{\verb{h}}{Hue}
\item{\verb{s}}{Saturation}
\item{\verb{r}}{Return value for the red component}
\item{\verb{g}}{Return value for the green component}
\item{\verb{b}}{Return value for the blue component}
}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
