\alias{gtkIconFactoryRemoveDefault}
\name{gtkIconFactoryRemoveDefault}
\title{gtkIconFactoryRemoveDefault}
\description{Removes an icon factory from the list of default icon
factories. Not normally used; you might use it for a library that
can be unloaded or shut down.}
\usage{gtkIconFactoryRemoveDefault(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconFactory}} previously added with \code{\link{gtkIconFactoryAddDefault}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
