\alias{gtkIconInfoCopy}
\name{gtkIconInfoCopy}
\title{gtkIconInfoCopy}
\description{Make a copy of a \verb{\link{GtkIconInfo}}.}
\usage{gtkIconInfoCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkIconInfo}}]  a \verb{\link{GtkIconInfo}}}}
\details{  Since  2.4}
\value{[\verb{\link{GtkIconInfo}}]  the new GtkIconInfo}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
