\alias{gtkIconInfoNewForPixbuf}
\name{gtkIconInfoNewForPixbuf}
\title{gtkIconInfoNewForPixbuf}
\description{Creates a \code{\link{GtkIconInfo}} for a \code{\link{GdkPixbuf}}.}
\usage{gtkIconInfoNewForPixbuf(icon.theme, pixbuf)}
\arguments{
\item{\verb{icon.theme}}{a \code{\link{GtkIconTheme}}}
\item{\verb{pixbuf}}{the pixbuf to wrap in a \code{\link{GtkIconInfo}}}
}
\details{Since 2.14}
\value{[\code{\link{GtkIconInfo}}] a \code{\link{GtkIconInfo}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
