\alias{gtkIconSetCopy}
\name{gtkIconSetCopy}
\title{gtkIconSetCopy}
\description{Copies \code{icon.set} by value.}
\usage{gtkIconSetCopy(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkIconSet}}]  a \code{\link{GtkIconSet}}}}
\value{[\code{\link{GtkIconSet}}]  a new \code{\link{GtkIconSet}} identical to the first.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
