\alias{gtkIconSetGetSizes}
\name{gtkIconSetGetSizes}
\title{gtkIconSetGetSizes}
\description{Obtains a list of icon sizes this icon set can render.}
\usage{gtkIconSetGetSizes(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkIconSet}}]  a \code{\link{GtkIconSet}}}}
\value{
A list containing the following elements:
\item{\code{sizes}}{[\code{\link{GtkIconSize}}]  return location for list of sizes}
\item{\code{n.sizes}}{[integer]  location to store number of elements in returned list}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
