\alias{gtkIconSizeFromName}
\name{gtkIconSizeFromName}
\title{gtkIconSizeFromName}
\description{Looks up the icon size associated with \code{name}.}
\usage{gtkIconSizeFromName(name)}
\arguments{\item{\verb{name}}{the name to look up.}}
\value{[\code{\link{GtkIconSize}}] the icon size with the given name.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
