\alias{gtkIconSourceGetDirection}
\name{gtkIconSourceGetDirection}
\title{gtkIconSourceGetDirection}
\description{Obtains the text direction this icon source applies to. The return
value is only useful/meaningful if the text direction is \emph{not}
wildcarded.}
\usage{gtkIconSourceGetDirection(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconSource}}}}
\value{[\code{\link{GtkTextDirection}}] text direction this source matches}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
