\alias{gtkIconSourceGetSize}
\name{gtkIconSourceGetSize}
\title{gtkIconSourceGetSize}
\description{Obtains the icon size this source applies to. The return value
is only useful/meaningful if the icon size is \emph{not} wildcarded.}
\usage{gtkIconSourceGetSize(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkIconSource}}]  a \verb{\link{GtkIconSource}}}}
\value{[\code{\link{GtkIconSize}}]  icon size this source matches.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
