\alias{gtkIconSourceGetSizeWildcarded}
\name{gtkIconSourceGetSizeWildcarded}
\title{gtkIconSourceGetSizeWildcarded}
\description{Gets the value set by \code{\link{gtkIconSourceSetSizeWildcarded}}.}
\usage{gtkIconSourceGetSizeWildcarded(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkIconSource}}]  a \verb{\link{GtkIconSource}}}}
\value{[logical] \code{TRUE} if this icon source is a base for any icon size variant}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
