\alias{gtkIconSourceGetState}
\name{gtkIconSourceGetState}
\title{gtkIconSourceGetState}
\description{Obtains the widget state this icon source applies to. The return
value is only useful/meaningful if the widget state is \emph{not}
wildcarded.}
\usage{gtkIconSourceGetState(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconSource}}}}
\value{[\code{\link{GtkStateType}}] widget state this source matches}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
