\alias{gtkIconSourceGetStateWildcarded}
\name{gtkIconSourceGetStateWildcarded}
\title{gtkIconSourceGetStateWildcarded}
\description{Gets the value set by \code{\link{gtkIconSourceSetStateWildcarded}}.}
\usage{gtkIconSourceGetStateWildcarded(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconSource}}}}
\value{[logical] \code{TRUE} if this icon source is a base for any widget state variant}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
