\alias{gtkIconSourceSetIconName}
\name{gtkIconSourceSetIconName}
\title{gtkIconSourceSetIconName}
\description{Sets the name of an icon to look up in the current icon theme
to use as a base image when creating icon variants for \verb{\link{GtkIconSet}}.}
\usage{gtkIconSourceSetIconName(object, icon.name)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkIconSource}}]  a \verb{\link{GtkIconSource}}}
\item{\verb{icon.name}}{[character]  name of icon to use}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
