\alias{gtkIconSourceSetPixbuf}
\name{gtkIconSourceSetPixbuf}
\title{gtkIconSourceSetPixbuf}
\description{Sets a pixbuf to use as a base image when creating icon variants
for \verb{\link{GtkIconSet}}.}
\usage{gtkIconSourceSetPixbuf(object, pixbuf)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkIconSource}}]  a \verb{\link{GtkIconSource}}}
\item{\verb{pixbuf}}{[\verb{\link{GdkPixbuf}}]  pixbuf to use as a source}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
