\alias{gtkIconThemeAppendSearchPath}
\name{gtkIconThemeAppendSearchPath}
\title{gtkIconThemeAppendSearchPath}
\description{Appends a directory to the search path. 
See \code{\link{gtkIconThemeSetSearchPath}}.}
\usage{gtkIconThemeAppendSearchPath(object, path)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkIconTheme}}}
\item{\verb{path}}{directory name to append to the icon path}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
