\alias{gtkIconThemeGetDefault}
\name{gtkIconThemeGetDefault}
\title{gtkIconThemeGetDefault}
\description{Gets the icon theme for the default screen. See
\code{\link{gtkIconThemeGetForScreen}}.}
\usage{gtkIconThemeGetDefault()}
\details{  Since  2.4}
\value{[\verb{\link{GtkIconTheme}}]  A unique \verb{\link{GtkIconTheme}} associated with
 the default screen. This icon theme is associated with
 the screen and can be used as long as the screen
 is open. Do not ref or unref it.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
