\alias{gtkIconThemeHasIcon}
\name{gtkIconThemeHasIcon}
\title{gtkIconThemeHasIcon}
\description{Checks whether an icon theme includes an icon
for a particular name.}
\usage{gtkIconThemeHasIcon(object, icon.name)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkIconTheme}}}
\item{\verb{icon.name}}{the name of an icon}
}
\details{Since 2.4}
\value{[logical] \code{TRUE} if \code{icon.theme} includes an
icon for \code{icon.name}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
