\alias{gtkIconThemeNew}
\name{gtkIconThemeNew}
\title{gtkIconThemeNew}
\description{Creates a new icon theme object. Icon theme objects are used
to lookup up an icon by name in a particular icon theme.
Usually, you'll want to use \code{\link{gtkIconThemeGetDefault}}
or \code{\link{gtkIconThemeGetForScreen}} rather than creating
a new icon theme object for scratch.}
\usage{gtkIconThemeNew()}
\details{Since 2.4}
\value{[\code{\link{GtkIconTheme}}] the newly created \code{\link{GtkIconTheme}} object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
