\alias{gtkIconThemePrependSearchPath}
\name{gtkIconThemePrependSearchPath}
\title{gtkIconThemePrependSearchPath}
\description{Prepends a directory to the search path. 
See \code{\link{gtkIconThemeSetSearchPath}}.}
\usage{gtkIconThemePrependSearchPath(object, path)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkIconTheme}}]  a \verb{\link{GtkIconTheme}}}
\item{\verb{path}}{[character]  directory name to prepend to the icon path}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
