\alias{gtkIconViewGetColumns}
\name{gtkIconViewGetColumns}
\title{gtkIconViewGetColumns}
\description{Returns the value of the ::columns property.}
\usage{gtkIconViewGetColumns(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconView}}}}
\details{Since 2.6}
\value{[integer] the number of columns, or -1}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
