\alias{gtkIconViewGetColumnSpacing}
\name{gtkIconViewGetColumnSpacing}
\title{gtkIconViewGetColumnSpacing}
\description{Returns the value of the ::column-spacing property.}
\usage{gtkIconViewGetColumnSpacing(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkIconView}}]  a \code{\link{GtkIconView}}}}
\details{  Since  2.6}
\value{[integer]  the space between columns}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
