\alias{gtkIconViewGetItemPadding}
\name{gtkIconViewGetItemPadding}
\title{gtkIconViewGetItemPadding}
\description{Returns the value of the ::item-padding property.}
\usage{gtkIconViewGetItemPadding(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconView}}}}
\details{Since 2.18}
\value{[integer] the padding around items}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
