\alias{gtkIconViewGetItemWidth}
\name{gtkIconViewGetItemWidth}
\title{gtkIconViewGetItemWidth}
\description{Returns the value of the ::item-width property.}
\usage{gtkIconViewGetItemWidth(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconView}}}}
\details{Since 2.6}
\value{[integer] the width of a single item, or -1}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
