\alias{gtkIconViewGetMargin}
\name{gtkIconViewGetMargin}
\title{gtkIconViewGetMargin}
\description{Returns the value of the ::margin property.}
\usage{gtkIconViewGetMargin(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconView}}}}
\details{Since 2.6}
\value{[integer] the space at the borders}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
