\alias{gtkIconViewGetMarkupColumn}
\name{gtkIconViewGetMarkupColumn}
\title{gtkIconViewGetMarkupColumn}
\description{Returns the column with markup text for \code{icon.view}.}
\usage{gtkIconViewGetMarkupColumn(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkIconView}}.}}
\details{Since 2.6}
\value{[integer] the markup column, or -1 if it's unset.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
