\alias{gtkIconViewGetModel}
\name{gtkIconViewGetModel}
\title{gtkIconViewGetModel}
\description{Returns the model the \verb{\link{GtkIconView}} is based on.  Returns \code{NULL} if the
model is unset.}
\usage{gtkIconViewGetModel(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkIconView}}]  a \verb{\link{GtkIconView}}}}
\details{  Since  2.6 }
\value{[\verb{\link{GtkTreeModel}}]  A \verb{\link{GtkTreeModel}}, or \code{NULL} if none is currently being used.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
