\alias{gtkIconViewGetOrientation}
\name{gtkIconViewGetOrientation}
\title{gtkIconViewGetOrientation}
\description{Returns the value of the ::orientation property which determines 
whether the labels are drawn beside the icons instead of below.}
\usage{gtkIconViewGetOrientation(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconView}}}}
\details{Since 2.6}
\value{[\code{\link{GtkOrientation}}] the relative position of texts and icons}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
