\alias{gtkIconViewGetPixbufColumn}
\name{gtkIconViewGetPixbufColumn}
\title{gtkIconViewGetPixbufColumn}
\description{Returns the column with pixbufs for \code{icon.view}.}
\usage{gtkIconViewGetPixbufColumn(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkIconView}}.}}
\details{Since 2.6}
\value{[integer] the pixbuf column, or -1 if it's unset.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
