\alias{gtkIconViewGetReorderable}
\name{gtkIconViewGetReorderable}
\title{gtkIconViewGetReorderable}
\description{Retrieves whether the user can reorder the list via drag-and-drop. 
See \code{\link{gtkIconViewSetReorderable}}.}
\usage{gtkIconViewGetReorderable(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconView}}}}
\details{Since 2.8}
\value{[logical] \code{TRUE} if the list can be reordered.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
