\alias{gtkIconViewGetSpacing}
\name{gtkIconViewGetSpacing}
\title{gtkIconViewGetSpacing}
\description{Returns the value of the ::spacing property.}
\usage{gtkIconViewGetSpacing(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkIconView}}]  a \verb{\link{GtkIconView}}}}
\details{  Since  2.6}
\value{[integer]  the space between cells }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
