\alias{gtkIconViewGetTextColumn}
\name{gtkIconViewGetTextColumn}
\title{gtkIconViewGetTextColumn}
\description{Returns the column with text for \code{icon.view}.}
\usage{gtkIconViewGetTextColumn(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkIconView}}.}}
\details{Since 2.6}
\value{[integer] the text column, or -1 if it's unset.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
