\alias{gtkIconViewGetTooltipColumn}
\name{gtkIconViewGetTooltipColumn}
\title{gtkIconViewGetTooltipColumn}
\description{Returns the column of \code{icon.view}'s model which is being used for
displaying tooltips on \code{icon.view}'s rows.}
\usage{gtkIconViewGetTooltipColumn(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconView}}}}
\details{Since 2.12}
\value{[integer] the index of the tooltip column that is currently being
used, or -1 if this is disabled.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
