\alias{gtkIconViewItemActivated}
\name{gtkIconViewItemActivated}
\title{gtkIconViewItemActivated}
\description{Activates the item determined by \code{path}.}
\usage{gtkIconViewItemActivated(object, path)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkIconView}}]  A \verb{\link{GtkIconView}}}
\item{\verb{path}}{[\verb{\link{GtkTreePath}}]  The \verb{\link{GtkTreePath}} to be activated}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
