\alias{gtkIconViewNew}
\name{gtkIconViewNew}
\title{gtkIconViewNew}
\description{Creates a new \verb{\link{GtkIconView}} widget}
\usage{gtkIconViewNew(show = TRUE)}
\details{  Since  2.6}
\value{[\verb{\link{GtkWidget}}]  A newly created \verb{\link{GtkIconView}} widget}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
