\alias{gtkIconViewNewWithModel}
\name{gtkIconViewNewWithModel}
\title{gtkIconViewNewWithModel}
\description{Creates a new \verb{\link{GtkIconView}} widget with the model \code{model}.}
\usage{gtkIconViewNewWithModel(model = NULL, show = TRUE)}
\arguments{\item{\verb{model}}{[\verb{\link{GtkTreeModel}}]  The model.}}
\details{  Since  2.6 }
\value{[\verb{\link{GtkWidget}}]  A newly created \verb{\link{GtkIconView}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
