\alias{gtkIconViewSelectPath}
\name{gtkIconViewSelectPath}
\title{gtkIconViewSelectPath}
\description{Selects the row at \code{path}.}
\usage{gtkIconViewSelectPath(object, path)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkIconView}}]  A \verb{\link{GtkIconView}}.}
\item{\verb{path}}{[\verb{\link{GtkTreePath}}]  The \verb{\link{GtkTreePath}} to be selected.}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
