\alias{gtkIconViewSetColumnSpacing}
\name{gtkIconViewSetColumnSpacing}
\title{gtkIconViewSetColumnSpacing}
\description{Sets the ::column-spacing property which specifies the space 
which is inserted between the columns of the icon view.}
\usage{gtkIconViewSetColumnSpacing(object, column.spacing)}
\arguments{
\item{\code{object}}{[\code{\link{GtkIconView}}]  a \code{\link{GtkIconView}}}
\item{\code{column.spacing}}{[integer]  the column spacing}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
