\alias{gtkIconViewSetDragDestItem}
\name{gtkIconViewSetDragDestItem}
\title{gtkIconViewSetDragDestItem}
\description{Sets the item that is highlighted for feedback.}
\usage{gtkIconViewSetDragDestItem(object, path, pos)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkIconView}}]  a \verb{\link{GtkIconView}}}
\item{\verb{path}}{[\verb{\link{GtkTreePath}}]  The path of the item to highlight, or \code{NULL}.}
\item{\verb{pos}}{[\code{\link{GtkIconViewDropPosition}}]  Specifies where to drop, relative to the item}
}
\details{  Since  2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
