\alias{gtkIconViewSetItemPadding}
\name{gtkIconViewSetItemPadding}
\title{gtkIconViewSetItemPadding}
\description{Sets the \verb{"item-padding"} property which specifies the padding
around each of the icon view's items.}
\usage{gtkIconViewSetItemPadding(object, item.padding)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkIconView}}}
\item{\verb{item.padding}}{the item padding}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
