\alias{gtkIconViewSetItemWidth}
\name{gtkIconViewSetItemWidth}
\title{gtkIconViewSetItemWidth}
\description{Sets the ::item-width property which specifies the width 
to use for each item. If it is set to -1, the icon view will 
automatically determine a suitable item size.}
\usage{gtkIconViewSetItemWidth(object, item.width)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkIconView}}}
\item{\verb{item.width}}{the width for each item}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
