\alias{gtkIconViewSetPixbufColumn}
\name{gtkIconViewSetPixbufColumn}
\title{gtkIconViewSetPixbufColumn}
\description{Sets the column with pixbufs for \code{icon.view} to be \code{column}. The pixbuf
column must be of type \verb{GDK_TYPE_PIXBUF}}
\usage{gtkIconViewSetPixbufColumn(object, column)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkIconView}}]  A \verb{\link{GtkIconView}}.}
\item{\verb{column}}{[integer]  A column in the currently used model, or -1 to disable}
}
\details{  Since  2.6 }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
