\alias{gtkIconViewSetRowSpacing}
\name{gtkIconViewSetRowSpacing}
\title{gtkIconViewSetRowSpacing}
\description{Sets the ::row-spacing property which specifies the space 
which is inserted between the rows of the icon view.}
\usage{gtkIconViewSetRowSpacing(object, row.spacing)}
\arguments{
\item{\code{object}}{[\code{\link{GtkIconView}}]  a \code{\link{GtkIconView}}}
\item{\code{row.spacing}}{[integer]  the row spacing}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
