\alias{gtkIconViewSetSelectionMode}
\name{gtkIconViewSetSelectionMode}
\title{gtkIconViewSetSelectionMode}
\description{Sets the selection mode of the \code{icon.view}.}
\usage{gtkIconViewSetSelectionMode(object, mode)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkIconView}}]  A \verb{\link{GtkIconView}}.}
\item{\verb{mode}}{[\code{\link{GtkSelectionMode}}]  The selection mode}
}
\details{  Since  2.6 }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
