\alias{gtkIconViewSetSpacing}
\name{gtkIconViewSetSpacing}
\title{gtkIconViewSetSpacing}
\description{Sets the ::spacing property which specifies the space 
which is inserted between the cells (i.e. the icon and 
the text) of an item.}
\usage{gtkIconViewSetSpacing(object, spacing)}
\arguments{
\item{\code{object}}{[\code{\link{GtkIconView}}]  a \code{\link{GtkIconView}}}
\item{\code{spacing}}{[integer]  the spacing}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
