\alias{gtkIconViewSetTooltipCell}
\name{gtkIconViewSetTooltipCell}
\title{gtkIconViewSetTooltipCell}
\description{Sets the tip area of \code{tooltip} to the area which \code{cell} occupies in
the item pointed to by \code{path}. See also \code{\link{gtkTooltipSetTipArea}}.}
\usage{gtkIconViewSetTooltipCell(object, tooltip, path, cell)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkIconView}}]  a \verb{\link{GtkIconView}}}
\item{\verb{tooltip}}{[\verb{\link{GtkTooltip}}]  a \verb{\link{GtkTooltip}}}
\item{\verb{path}}{[\verb{\link{GtkTreePath}}]  a \verb{\link{GtkTreePath}}}
\item{\verb{cell}}{[\verb{\link{GtkCellRenderer}}]  a \verb{\link{GtkCellRenderer}} or \code{NULL}}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
