\alias{gtkIconViewSetTooltipItem}
\name{gtkIconViewSetTooltipItem}
\title{gtkIconViewSetTooltipItem}
\description{Sets the tip area of \code{tooltip} to be the area covered by the item at \code{path}.
See also \code{\link{gtkTooltipSetTipArea}}.}
\usage{gtkIconViewSetTooltipItem(object, tooltip, path)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkIconView}}]  a \verb{\link{GtkIconView}}}
\item{\verb{tooltip}}{[\verb{\link{GtkTooltip}}]  a \verb{\link{GtkTooltip}}}
\item{\verb{path}}{[\verb{\link{GtkTreePath}}]  a \verb{\link{GtkTreePath}}}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
