\alias{gtkIconViewUnselectPath}
\name{gtkIconViewUnselectPath}
\title{gtkIconViewUnselectPath}
\description{Unselects the row at \code{path}.}
\usage{gtkIconViewUnselectPath(object, path)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkIconView}}]  A \verb{\link{GtkIconView}}.}
\item{\verb{path}}{[\verb{\link{GtkTreePath}}]  The \verb{\link{GtkTreePath}} to be unselected.}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
