\alias{gtkIconViewUnsetModelDragDest}
\name{gtkIconViewUnsetModelDragDest}
\title{gtkIconViewUnsetModelDragDest}
\description{Undoes the effect of \code{\link{gtkIconViewEnableModelDragDest}}.}
\usage{gtkIconViewUnsetModelDragDest(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkIconView}}]  a \verb{\link{GtkIconView}}}}
\details{  Since  2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
