\alias{gtkIconViewUnsetModelDragSource}
\name{gtkIconViewUnsetModelDragSource}
\title{gtkIconViewUnsetModelDragSource}
\description{Undoes the effect of \code{\link{gtkIconViewEnableModelDragSource}}.}
\usage{gtkIconViewUnsetModelDragSource(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkIconView}}]  a \verb{\link{GtkIconView}}}}
\details{  Since  2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
