\alias{gtkImageGetStorageType}
\name{gtkImageGetStorageType}
\title{gtkImageGetStorageType}
\description{Gets the type of representation being used by the \verb{\link{GtkImage}}
to store image data. If the \verb{\link{GtkImage}} has no image data,
the return value will be \code{GTK_IMAGE_EMPTY}.}
\usage{gtkImageGetStorageType(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkImage}}]  a \verb{\link{GtkImage}}}}
\value{[\code{\link{GtkImageType}}]  image representation being used}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
