\alias{gtkImageMenuItemGetAlwaysShowImage}
\name{gtkImageMenuItemGetAlwaysShowImage}
\title{gtkImageMenuItemGetAlwaysShowImage}
\description{Returns whether the menu item will ignore the \verb{"gtk-menu-images"}
setting and always show the image, if available.}
\usage{gtkImageMenuItemGetAlwaysShowImage(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkImageMenuItem}}}}
\details{Since 2.16}
\value{[logical] \code{TRUE} if the menu item will always show the image}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
