\alias{gtkImageMenuItemGetImage}
\name{gtkImageMenuItemGetImage}
\title{gtkImageMenuItemGetImage}
\description{Gets the widget that is currently set as the image of \code{image.menu.item}.
See \code{\link{gtkImageMenuItemSetImage}}.}
\usage{gtkImageMenuItemGetImage(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkImageMenuItem}}.}}
\value{[\code{\link{GtkWidget}}] the widget set as image of \code{image.menu.item}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
