\alias{gtkImageMenuItemGetUseStock}
\name{gtkImageMenuItemGetUseStock}
\title{gtkImageMenuItemGetUseStock}
\description{Checks whether the label set in the menuitem is used as a
stock id to select the stock item for the item.}
\usage{gtkImageMenuItemGetUseStock(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkImageMenuItem}}}}
\details{Since 2.16}
\value{[logical] \code{TRUE} if the label set in the menuitem is used as a
stock id to select the stock item for the item}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
