\alias{gtkImageMenuItemNew}
\name{gtkImageMenuItemNew}
\title{gtkImageMenuItemNew}
\description{Creates a new \verb{\link{GtkImageMenuItem}} with an empty label.}
\usage{gtkImageMenuItemNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}]  a new \verb{\link{GtkImageMenuItem}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
