\alias{gtkImageMenuItemSetImage}
\name{gtkImageMenuItemSetImage}
\title{gtkImageMenuItemSetImage}
\description{Sets the image of \code{image.menu.item} to the given widget.
Note that it depends on the show-menu-images setting whether
the image will be displayed or not.}
\usage{gtkImageMenuItemSetImage(object, image = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkImageMenuItem}}.}
\item{\verb{image}}{a widget to set as the image for the menu item. \emph{[  \acronym{allow-none}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
