\alias{gtkImageMenuItemSetUseStock}
\name{gtkImageMenuItemSetUseStock}
\title{gtkImageMenuItemSetUseStock}
\description{If \code{TRUE}, the label set in the menuitem is used as a
stock id to select the stock item for the item.}
\usage{gtkImageMenuItemSetUseStock(object, use.stock)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkImageMenuItem}}}
\item{\verb{use.stock}}{\code{TRUE} if the menuitem should use a stock item}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
