\alias{gtkImageSet}
\name{gtkImageSet}
\title{gtkImageSet}
\description{
Sets the \code{\link{GtkImage}}.
\strong{WARNING: \code{gtk_image_set} is deprecated and should not be used in newly-written code.}
}
\usage{gtkImageSet(object, val, mask)}
\arguments{
\item{\code{object}}{[\code{\link{GtkImage}}] a \code{\link{GtkImage}}}
\item{\code{val}}{[\code{\link{GdkImage}}] a \code{\link{GdkImage}}}
\item{\code{mask}}{[\code{\link{GdkBitmap}}] a \code{\link{GdkBitmap}} that indicates which parts of the image should be transparent.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
