\alias{gtkImageSetFromAnimation}
\name{gtkImageSetFromAnimation}
\title{gtkImageSetFromAnimation}
\description{Causes the \verb{\link{GtkImage}} to display the given animation (or display
nothing, if you set the animation to \code{NULL}).}
\usage{gtkImageSetFromAnimation(object, animation)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkImage}}]  a \verb{\link{GtkImage}}}
\item{\verb{animation}}{[\verb{\link{GdkPixbufAnimation}}]  the \verb{\link{GdkPixbufAnimation}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
