\alias{gtkImageSetFromFile}
\name{gtkImageSetFromFile}
\title{gtkImageSetFromFile}
\description{See \code{\link{gtkImageNewFromFile}} for details.}
\usage{gtkImageSetFromFile(object, filename = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkImage}}]  a \verb{\link{GtkImage}}}
\item{\verb{filename}}{[character]  a filename or \code{NULL}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
