\alias{gtkImageSetFromGicon}
\name{gtkImageSetFromGicon}
\title{gtkImageSetFromGicon}
\description{See \code{\link{gtkImageNewFromGicon}} for details.}
\usage{gtkImageSetFromGicon(object, icon, size)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkImage}}}
\item{\verb{icon}}{an icon}
\item{\verb{size}}{an icon size. \emph{[  \acronym{type}   int]}}
}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
