\alias{gtkImageSetFromIconName}
\name{gtkImageSetFromIconName}
\title{gtkImageSetFromIconName}
\description{See \code{\link{gtkImageNewFromIconName}} for details.}
\usage{gtkImageSetFromIconName(object, icon.name, size)}
\arguments{
\item{\code{object}}{[\code{\link{GtkImage}}]  a \code{\link{GtkImage}}}
\item{\code{icon.name}}{[character]  an icon name}
\item{\code{size}}{[\code{\link{GtkIconSize}}]  an icon size}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
