\alias{gtkImageSetFromIconSet}
\name{gtkImageSetFromIconSet}
\title{gtkImageSetFromIconSet}
\description{See \code{\link{gtkImageNewFromIconSet}} for details.}
\usage{gtkImageSetFromIconSet(object, icon.set, size)}
\arguments{
\item{\code{object}}{[\code{\link{GtkImage}}]  a \code{\link{GtkImage}}}
\item{\code{icon.set}}{[\code{\link{GtkIconSet}}]  a \code{\link{GtkIconSet}}}
\item{\code{size}}{[\code{\link{GtkIconSize}}]  a stock icon size}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
