\alias{gtkImageSetFromImage}
\name{gtkImageSetFromImage}
\title{gtkImageSetFromImage}
\description{See \code{\link{gtkImageNewFromImage}} for details.}
\usage{gtkImageSetFromImage(object, gdk.image = NULL, mask = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkImage}}]  a \code{\link{GtkImage}}}
\item{\code{gdk.image}}{[\code{\link{GdkImage}}]  a \code{\link{GdkImage}} or \code{NULL}}
\item{\code{mask}}{[\code{\link{GdkBitmap}}]  a \code{\link{GdkBitmap}} or \code{NULL}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
