\alias{gtkImageSetFromPixmap}
\name{gtkImageSetFromPixmap}
\title{gtkImageSetFromPixmap}
\description{See \code{\link{gtkImageNewFromPixmap}} for details.}
\usage{gtkImageSetFromPixmap(object, pixmap, mask = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkImage}}]  a \verb{\link{GtkImage}}}
\item{\verb{pixmap}}{[\verb{\link{GdkPixmap}}]  a \verb{\link{GdkPixmap}} or \code{NULL}}
\item{\verb{mask}}{[\verb{\link{GdkBitmap}}]  a \verb{\link{GdkBitmap}} or \code{NULL}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
