\alias{gtkImageSetFromStock}
\name{gtkImageSetFromStock}
\title{gtkImageSetFromStock}
\description{See \code{\link{gtkImageNewFromStock}} for details.}
\usage{gtkImageSetFromStock(object, stock.id, size)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkImage}}]  a \verb{\link{GtkImage}}}
\item{\verb{stock.id}}{[character]  a stock icon name}
\item{\verb{size}}{[\code{\link{GtkIconSize}}]  a stock icon size}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
