\alias{gtkImageSetPixelSize}
\name{gtkImageSetPixelSize}
\title{gtkImageSetPixelSize}
\description{Sets the pixel size to use for named icons. If the pixel size is set
to a value != -1, it is used instead of the icon size set by
\code{\link{gtkImageSetFromIconName}}.}
\usage{gtkImageSetPixelSize(object, pixel.size)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkImage}}}
\item{\verb{pixel.size}}{the new pixel size}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
