\alias{gtkIMContextFocusIn}
\name{gtkIMContextFocusIn}
\title{gtkIMContextFocusIn}
\description{Notify the input method that the widget to which this
input context corresponds has gained focus. The input method
may, for example, change the displayed feedback to reflect
this change.}
\usage{gtkIMContextFocusIn(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIMContext}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
