\alias{gtkIMContextFocusOut}
\name{gtkIMContextFocusOut}
\title{gtkIMContextFocusOut}
\description{Notify the input method that the widget to which this
input context corresponds has lost focus. The input method
may, for example, change the displayed feedback or reset the contexts
state to reflect this change.}
\usage{gtkIMContextFocusOut(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIMContext}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
