\alias{gtkIMContextReset}
\name{gtkIMContextReset}
\title{gtkIMContextReset}
\description{Notify the input method that a change such as a change in cursor
position has been made. This will typically cause the input
method to clear the preedit state.}
\usage{gtkIMContextReset(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIMContext}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
