\alias{gtkIMContextSetCursorLocation}
\name{gtkIMContextSetCursorLocation}
\title{gtkIMContextSetCursorLocation}
\description{Notify the input method that a change in cursor 
position has been made. The location is relative to the client
window.}
\usage{gtkIMContextSetCursorLocation(object, area)}
\arguments{
\item{\code{object}}{[\code{\link{GtkIMContext}}]  a \code{\link{GtkIMContext}}}
\item{\code{area}}{[\code{\link{GdkRectangle}}]  new location}
}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
