\alias{gtkIMContextSimpleNew}
\name{gtkIMContextSimpleNew}
\title{gtkIMContextSimpleNew}
\description{Creates a new \code{\link{GtkIMContextSimple}}.}
\usage{gtkIMContextSimpleNew()}
\value{[\code{\link{GtkIMContext}}]  a new \code{\link{GtkIMContextSimple}}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
