\alias{gtkIMMulticontextAppendMenuitems}
\name{gtkIMMulticontextAppendMenuitems}
\title{gtkIMMulticontextAppendMenuitems}
\description{Add menuitems for various available input methods to a menu;
the menuitems, when selected, will switch the input method
for the context and the global default input method.}
\usage{gtkIMMulticontextAppendMenuitems(object, menushell)}
\arguments{
\item{\code{object}}{[\code{\link{GtkIMMulticontext}}]  a \code{GtkIMMultiContext}}
\item{\code{menushell}}{[\code{\link{GtkMenuShell}}]  a \code{\link{GtkMenuShell}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
