\alias{gtkIMMulticontextGetContextId}
\name{gtkIMMulticontextGetContextId}
\title{gtkIMMulticontextGetContextId}
\description{Gets the id of the currently active slave of the \code{context}.}
\usage{gtkIMMulticontextGetContextId(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIMMulticontext}}}}
\details{Since 2.16}
\value{[char] the id of the currently active slave}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
