\alias{gtkIMMulticontextNew}
\name{gtkIMMulticontextNew}
\title{gtkIMMulticontextNew}
\description{Creates a new \verb{\link{GtkIMMulticontext}}.}
\usage{gtkIMMulticontextNew()}
\value{[\verb{\link{GtkIMContext}}]  a new \verb{\link{GtkIMMulticontext}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
