\alias{gtkInfoBarGetActionArea}
\name{gtkInfoBarGetActionArea}
\title{gtkInfoBarGetActionArea}
\description{Returns the action area of \code{info.bar}.}
\usage{gtkInfoBarGetActionArea(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkInfoBar}}}}
\details{Since 2.18}
\value{[\code{\link{GtkWidget}}] the action area.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
