\alias{gtkInfoBarGetContentArea}
\name{gtkInfoBarGetContentArea}
\title{gtkInfoBarGetContentArea}
\description{Returns the content area of \code{info.bar}.}
\usage{gtkInfoBarGetContentArea(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkInfoBar}}}}
\details{Since 2.18}
\value{[\code{\link{GtkWidget}}] the content area.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
