\alias{gtkInfoBarGetMessageType}
\name{gtkInfoBarGetMessageType}
\title{gtkInfoBarGetMessageType}
\description{Returns the message type of the message area.}
\usage{gtkInfoBarGetMessageType(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkInfoBar}}}}
\details{Since 2.18}
\value{[\code{\link{GtkMessageType}}] the message type of the message area.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
