\alias{gtkInfoBarNew}
\name{gtkInfoBarNew}
\title{gtkInfoBarNew}
\description{Creates a new \code{\link{GtkInfoBar}} object.}
\usage{gtkInfoBarNew(show = TRUE)}
\details{Since 2.18}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkInfoBar}} object}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
