\alias{gtkInfoBarResponse}
\name{gtkInfoBarResponse}
\title{gtkInfoBarResponse}
\description{Emits the 'response' signal with the given \code{response.id}.}
\usage{gtkInfoBarResponse(object, response.id)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkInfoBar}}}
\item{\verb{response.id}}{a response ID}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
