\alias{gtkInfoBarSetMessageType}
\name{gtkInfoBarSetMessageType}
\title{gtkInfoBarSetMessageType}
\description{Sets the message type of the message area.
GTK+ uses this type to determine what color to use
when drawing the message area.}
\usage{gtkInfoBarSetMessageType(object, message.type)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkInfoBar}}}
\item{\verb{message.type}}{a \code{\link{GtkMessageType}}}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
