\alias{gtkInitAdd}
\name{gtkInitAdd}
\title{gtkInitAdd}
\description{Registers a function to be called when the mainloop is started.}
\usage{gtkInitAdd(fun, data = NULL)}
\arguments{\item{\code{data}}{[R object] Data to pass to that function.}}
\note{
This function is deprecated. You should use \code{GDK_PRIORITY_REDRAW} instead.
This function is deprecated. You should use \code{G_PRIORITY_HIGH} instead.
This function is deprecated. You should use \code{G_PRIORITY_DEFAULT_IDLE} instead.
This function is deprecated. You should use \code{G_PRIORITY_LOW} instead.
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
