\alias{gtkInputDialogNew}
\name{gtkInputDialogNew}
\title{gtkInputDialogNew}
\description{Creates a new \code{\link{GtkInputDialog}}.}
\usage{gtkInputDialogNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] the new \code{\link{GtkInputDialog}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
