\alias{gtkInvisibleGetScreen}
\name{gtkInvisibleGetScreen}
\title{gtkInvisibleGetScreen}
\description{Returns the \verb{\link{GdkScreen}} object associated with \code{invisible}}
\usage{gtkInvisibleGetScreen(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkInvisible}}]  a \verb{\link{GtkInvisible}}.}}
\details{  Since  2.2}
\value{[\verb{\link{GdkScreen}}]  the associated \verb{\link{GdkScreen}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
