\alias{gtkInvisibleNewForScreen}
\name{gtkInvisibleNewForScreen}
\title{gtkInvisibleNewForScreen}
\description{Creates a new \verb{\link{GtkInvisible}} object for a specified screen}
\usage{gtkInvisibleNewForScreen(screen, show = TRUE)}
\arguments{\item{\verb{screen}}{[\verb{\link{GdkScreen}}]  a \verb{\link{GdkScreen}} which identifies on which
    the new \verb{\link{GtkInvisible}} will be created.}}
\details{  Since  2.2}
\value{[\verb{\link{GtkWidget}}]  a newly created \verb{\link{GtkInvisible}} object}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
