\alias{gtkInvisibleSetScreen}
\name{gtkInvisibleSetScreen}
\title{gtkInvisibleSetScreen}
\description{Sets the \verb{\link{GdkScreen}} where the \verb{\link{GtkInvisible}} object will be displayed.}
\usage{gtkInvisibleSetScreen(object, screen)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkInvisible}}]  a \verb{\link{GtkInvisible}}.}
\item{\verb{screen}}{[\verb{\link{GdkScreen}}]  a \verb{\link{GdkScreen}}.}
}
\details{  Since  2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
