\alias{gtkItemDeselect}
\name{gtkItemDeselect}
\title{gtkItemDeselect}
\description{Emits the "deselect" signal on the given item.}
\usage{gtkItemDeselect(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkItem}}] a \code{\link{GtkItem}}.}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
