\alias{gtkItemFactoryCreateMenuEntries}
\name{gtkItemFactoryCreateMenuEntries}
\title{gtkItemFactoryCreateMenuEntries}
\description{
Creates the menu items from the \code{entries}.
\strong{WARNING: \code{gtk_item_factory_create_menu_entries} is deprecated and should not be used in newly-written code.}
}
\usage{gtkItemFactoryCreateMenuEntries(entries)}
\arguments{\item{\code{entries}}{[GtkMenuEntry]  a list of \code{GtkMenuEntry}s }}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
