\alias{gtkItemFactoryDeleteItem}
\name{gtkItemFactoryDeleteItem}
\title{gtkItemFactoryDeleteItem}
\description{
Deletes the menu item which was created for \code{path} by the given
item factory.
\strong{WARNING: \code{gtk_item_factory_delete_item} is deprecated and should not be used in newly-written code.}
}
\usage{gtkItemFactoryDeleteItem(object, path)}
\arguments{
\item{\code{object}}{[\code{\link{GtkItemFactory}}]  a \code{\link{GtkItemFactory}}}
\item{\code{path}}{[character]  a path}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
