\alias{gtkItemFactoryFromWidget}
\name{gtkItemFactoryFromWidget}
\title{gtkItemFactoryFromWidget}
\description{
Obtains the item factory from which a widget was created.
\strong{WARNING: \code{gtk_item_factory_from_widget} is deprecated and should not be used in newly-written code. 2.4}
}
\usage{gtkItemFactoryFromWidget(widget)}
\arguments{\item{\verb{widget}}{[\verb{\link{GtkWidget}}]  a widget}}
\value{[\verb{\link{GtkItemFactory}}]  the item factory from which \code{widget} was created, or \code{NULL}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
