\alias{gtkItemSelect}
\name{gtkItemSelect}
\title{gtkItemSelect}
\description{Emits the "select" signal on the given item.}
\usage{gtkItemSelect(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkItem}}] a \code{\link{GtkItem}}.}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
