\alias{gtkItemToggle}
\name{gtkItemToggle}
\title{gtkItemToggle}
\description{Emits the "toggle" signal on the given item.}
\usage{gtkItemToggle(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkItem}}] a \code{\link{GtkItem}}.}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
