\alias{gtkKeySnooperRemove}
\name{gtkKeySnooperRemove}
\title{gtkKeySnooperRemove}
\description{Removes the key snooper function with the given id.}
\usage{gtkKeySnooperRemove(snooper.handler.id)}
\arguments{\item{\verb{snooper.handler.id}}{Identifies the key snooper to remove.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
