\alias{gtkLabelGetEllipsize}
\name{gtkLabelGetEllipsize}
\title{gtkLabelGetEllipsize}
\description{Returns the ellipsizing position of the label. See \code{\link{gtkLabelSetEllipsize}}.}
\usage{gtkLabelGetEllipsize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkLabel}}}}
\details{Since 2.6}
\value{[\code{\link{PangoEllipsizeMode}}] \code{\link{PangoEllipsizeMode}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
