\alias{gtkLabelGetJustify}
\name{gtkLabelGetJustify}
\title{gtkLabelGetJustify}
\description{Returns the justification of the label. See \code{\link{gtkLabelSetJustify}}.}
\usage{gtkLabelGetJustify(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkLabel}}}}
\value{[\code{\link{GtkJustification}}] \code{\link{GtkJustification}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
