\alias{gtkLabelGetLabel}
\name{gtkLabelGetLabel}
\title{gtkLabelGetLabel}
\description{Fetches the text from a label widget including any embedded
underlines indicating mnemonics and Pango markup. (See
\code{\link{gtkLabelGetText}}).}
\usage{gtkLabelGetLabel(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkLabel}}]  a \code{\link{GtkLabel}}}}
\value{[character]  the text of the label widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
