\alias{gtkLabelGetLineWrap}
\name{gtkLabelGetLineWrap}
\title{gtkLabelGetLineWrap}
\description{Returns whether lines in the label are automatically wrapped. 
See \code{\link{gtkLabelSetLineWrap}}.}
\usage{gtkLabelGetLineWrap(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkLabel}}}}
\value{[logical] \code{TRUE} if the lines of the label are automatically wrapped.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
