\alias{gtkLabelGetMaxWidthChars}
\name{gtkLabelGetMaxWidthChars}
\title{gtkLabelGetMaxWidthChars}
\description{Retrieves the desired maximum width of \code{label}, in characters. See
\code{\link{gtkLabelSetWidthChars}}.}
\usage{gtkLabelGetMaxWidthChars(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkLabel}}}}
\details{Since 2.6}
\value{[integer] the maximum width of the label in characters.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
