\alias{gtkLabelGetMnemonicWidget}
\name{gtkLabelGetMnemonicWidget}
\title{gtkLabelGetMnemonicWidget}
\description{Retrieves the target of the mnemonic (keyboard shortcut) of this
label. See \code{\link{gtkLabelSetMnemonicWidget}}.}
\usage{gtkLabelGetMnemonicWidget(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkLabel}}}}
\value{[\code{\link{GtkWidget}}] the target of the label's mnemonic, or \code{NULL} if none
has been set and the default algorithm will be used.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
