\alias{gtkLabelGetSingleLineMode}
\name{gtkLabelGetSingleLineMode}
\title{gtkLabelGetSingleLineMode}
\description{Returns whether the label is in single line mode.}
\usage{gtkLabelGetSingleLineMode(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkLabel}}}}
\details{Since 2.6}
\value{[logical] \code{TRUE} when the label is in single line mode.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
