\alias{gtkLabelGetTrackVisitedLinks}
\name{gtkLabelGetTrackVisitedLinks}
\title{gtkLabelGetTrackVisitedLinks}
\description{Returns whether the label is currently keeping track
of clicked links.}
\usage{gtkLabelGetTrackVisitedLinks(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkLabel}}}}
\details{Since 2.18}
\value{[logical] \code{TRUE} if clicked links are remembered}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
