\alias{gtkLabelGetUseMarkup}
\name{gtkLabelGetUseMarkup}
\title{gtkLabelGetUseMarkup}
\description{Returns whether the label's text is interpreted as marked up with
the Pango text markup
language. See \code{\link{gtkLabelSetUseMarkup}}.}
\usage{gtkLabelGetUseMarkup(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkLabel}}}}
\value{[logical] \code{TRUE} if the label's text will be parsed for markup.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
