\alias{gtkLabelGetUseUnderline}
\name{gtkLabelGetUseUnderline}
\title{gtkLabelGetUseUnderline}
\description{Returns whether an embedded underline in the label indicates a
mnemonic. See \code{\link{gtkLabelSetUseUnderline}}.}
\usage{gtkLabelGetUseUnderline(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkLabel}}}}
\value{[logical] \code{TRUE} whether an embedded underline in the label indicates
the mnemonic accelerator keys.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
