\alias{gtkLabelGetWidthChars}
\name{gtkLabelGetWidthChars}
\title{gtkLabelGetWidthChars}
\description{Retrieves the desired width of \code{label}, in characters. See
\code{\link{gtkLabelSetWidthChars}}.}
\usage{gtkLabelGetWidthChars(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkLabel}}}}
\details{Since 2.6}
\value{[integer] the width of the label in characters.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
