\alias{gtkLabelNew}
\name{gtkLabelNew}
\title{gtkLabelNew}
\description{Creates a new label with the given text inside it. You can
pass \code{NULL} to get an empty label widget.}
\usage{gtkLabelNew(str = NULL, show = TRUE)}
\arguments{\item{\verb{str}}{The text of the label}}
\value{[\code{\link{GtkWidget}}] the new \code{\link{GtkLabel}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
