\alias{gtkLabelSetAngle}
\name{gtkLabelSetAngle}
\title{gtkLabelSetAngle}
\description{Sets the angle of rotation for the label. An angle of 90 reads from
from bottom to top, an angle of 270, from top to bottom. The angle
setting for the label is ignored if the label is selectable,
wrapped, or ellipsized.}
\usage{gtkLabelSetAngle(object, angle)}
\arguments{
\item{\code{object}}{[\code{\link{GtkLabel}}]  a \code{\link{GtkLabel}}}
\item{\code{angle}}{[numeric]  the angle that the baseline of the label makes with
  the horizontal, in degrees, measured counterclockwise}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
