\alias{gtkLabelSetLabel}
\name{gtkLabelSetLabel}
\title{gtkLabelSetLabel}
\description{Sets the text of the label. The label is interpreted as
including embedded underlines and/or Pango markup depending
on the values of label->use\_underline and label->use\_markup.}
\usage{gtkLabelSetLabel(object, str)}
\arguments{
\item{\code{object}}{[\code{\link{GtkLabel}}]  a \code{\link{GtkLabel}}}
\item{\code{str}}{[character]  the new text to set for the label}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
