\alias{gtkLabelSetLineWrap}
\name{gtkLabelSetLineWrap}
\title{gtkLabelSetLineWrap}
\description{Toggles line wrapping within the \code{\link{GtkLabel}} widget.  \code{TRUE} makes it break
lines if text exceeds the widget's size.  \code{FALSE} lets the text get cut off
by the edge of the widget if it exceeds the widget size.}
\usage{gtkLabelSetLineWrap(object, wrap)}
\arguments{
\item{\code{object}}{[\code{\link{GtkLabel}}]  a \code{\link{GtkLabel}}}
\item{\code{wrap}}{[logical]  the setting}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
