\alias{gtkLabelSetLineWrapMode}
\name{gtkLabelSetLineWrapMode}
\title{gtkLabelSetLineWrapMode}
\description{If line wrapping is on (see \code{\link{gtkLabelSetLineWrap}}) this controls how
the line wrapping is done. The default is \code{PANGO_WRAP_WORD} which means
wrap on word boundaries.}
\usage{gtkLabelSetLineWrapMode(object, wrap.mode)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkLabel}}}
\item{\verb{wrap.mode}}{the line wrapping mode}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
