\alias{gtkLabelSetMaxWidthChars}
\name{gtkLabelSetMaxWidthChars}
\title{gtkLabelSetMaxWidthChars}
\description{Sets the desired maximum width in characters of \code{label} to \code{n.chars}.}
\usage{gtkLabelSetMaxWidthChars(object, n.chars)}
\arguments{
\item{\code{object}}{[\code{\link{GtkLabel}}]  a \code{\link{GtkLabel}}}
\item{\code{n.chars}}{[integer]  the new desired maximum width, in characters.}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
