\alias{gtkLabelSetSingleLineMode}
\name{gtkLabelSetSingleLineMode}
\title{gtkLabelSetSingleLineMode}
\description{Sets whether the label is in single line mode.}
\usage{gtkLabelSetSingleLineMode(object, single.line.mode)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkLabel}}]  a \verb{\link{GtkLabel}}}
\item{\verb{single.line.mode}}{[logical] \code{TRUE} if the label should be in single line mode}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
