\alias{gtkLabelSetText}
\name{gtkLabelSetText}
\title{gtkLabelSetText}
\description{Sets the text within the \verb{\link{GtkLabel}} widget. It overwrites any text that
was there before.  }
\usage{gtkLabelSetText(object, str)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkLabel}}]  a \verb{\link{GtkLabel}}}
\item{\verb{str}}{[character]  The text you want to set}
}
\details{This will also clear any previously set mnemonic accelerators.  }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
