\alias{gtkLabelSetTrackVisitedLinks}
\name{gtkLabelSetTrackVisitedLinks}
\title{gtkLabelSetTrackVisitedLinks}
\description{Sets whether the label should keep track of clicked
links (and use a different color for them).}
\usage{gtkLabelSetTrackVisitedLinks(object, track.links)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkLabel}}}
\item{\verb{track.links}}{\code{TRUE} to track visited links}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
