\alias{gtkLayoutFreeze}
\name{gtkLayoutFreeze}
\title{gtkLayoutFreeze}
\description{
This is a deprecated function, it doesn't do anything useful.
\strong{WARNING: \code{gtk_layout_freeze} is deprecated and should not be used in newly-written code.  
This is a deprecated function, it doesn't do anything useful.  }
}
\usage{gtkLayoutFreeze(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkLayout}}]  a \code{\link{GtkLayout}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
