\alias{gtkLayoutGetBinWindow}
\name{gtkLayoutGetBinWindow}
\title{gtkLayoutGetBinWindow}
\description{Retrieve the bin window of the layout used for drawing operations.}
\usage{gtkLayoutGetBinWindow(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkLayout}}}}
\details{Since 2.14}
\value{[\code{\link{GdkWindow}}] a \code{\link{GdkWindow}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
