\alias{gtkLayoutMove}
\name{gtkLayoutMove}
\title{gtkLayoutMove}
\description{Moves a current child of \code{layout} to a new position.}
\usage{gtkLayoutMove(object, child.widget, x, y)}
\arguments{
\item{\code{object}}{[\code{\link{GtkLayout}}]  a \code{\link{GtkLayout}}}
\item{\code{child.widget}}{[\code{\link{GtkWidget}}]  a current child of \code{layout}}
\item{\code{x}}{[integer]  X position to move to}
\item{\code{y}}{[integer]  Y position to move to}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
