\alias{gtkLayoutPut}
\name{gtkLayoutPut}
\title{gtkLayoutPut}
\description{Adds \code{child.widget} to \code{layout}, at position (\code{x},\code{y}).
\code{layout} becomes the new parent container of \code{child.widget}.}
\usage{gtkLayoutPut(object, child.widget, x, y)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkLayout}}]  a \verb{\link{GtkLayout}}}
\item{\verb{child.widget}}{[\verb{\link{GtkWidget}}]  child widget}
\item{\verb{x}}{[integer]  X position of child widget}
\item{\verb{y}}{[integer]  Y position of child widget}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
