\alias{gtkLayoutSetSize}
\name{gtkLayoutSetSize}
\title{gtkLayoutSetSize}
\description{Sets the size of the scrollable area of the layout.}
\usage{gtkLayoutSetSize(object, width, height)}
\arguments{
\item{\code{object}}{[\code{\link{GtkLayout}}]  a \code{\link{GtkLayout}}}
\item{\code{width}}{[numeric]  width of entire scrollable area}
\item{\code{height}}{[numeric]  height of entire scrollable area}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
