\alias{gtkLayoutSetVadjustment}
\name{gtkLayoutSetVadjustment}
\title{gtkLayoutSetVadjustment}
\description{Sets the vertical scroll adjustment for the layout.}
\usage{gtkLayoutSetVadjustment(object, adjustment = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkLayout}}]  a \code{\link{GtkLayout}}}
\item{\code{adjustment}}{[\code{\link{GtkAdjustment}}]  new scroll adjustment}
}
\details{See \code{\link{GtkScrolledWindow}}, \code{\link{GtkScrollbar}}, \code{\link{GtkAdjustment}} for details.  }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
