\alias{gtkLinkButtonGetUri}
\name{gtkLinkButtonGetUri}
\title{gtkLinkButtonGetUri}
\description{Retrieves the URI set using \code{\link{gtkLinkButtonSetUri}}.}
\usage{gtkLinkButtonGetUri(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkLinkButton}}]  a \verb{\link{GtkLinkButton}}}}
\details{  Since  2.10}
\value{[character]  a valid URI.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
