\alias{gtkLinkButtonNew}
\name{gtkLinkButtonNew}
\title{gtkLinkButtonNew}
\description{Creates a new \verb{\link{GtkLinkButton}} with the URI as its text.}
\usage{gtkLinkButtonNew(uri)}
\arguments{\item{\verb{uri}}{[character]  a valid URI}}
\details{  Since  2.10}
\value{[\verb{\link{GtkWidget}}]  a new link button widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
