\alias{gtkLinkButtonNewWithLabel}
\name{gtkLinkButtonNewWithLabel}
\title{gtkLinkButtonNewWithLabel}
\description{Creates a new \verb{\link{GtkLinkButton}} containing a label.}
\usage{gtkLinkButtonNewWithLabel(uri, label = NULL, show = TRUE)}
\arguments{
\item{\verb{uri}}{[character]  a valid URI}
\item{\verb{label}}{[character]  the text of the button}
}
\details{  Since  2.10}
\value{[\verb{\link{GtkWidget}}]  a new link button widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
