\alias{gtkLinkButtonSetUri}
\name{gtkLinkButtonSetUri}
\title{gtkLinkButtonSetUri}
\description{Sets \code{uri} as the URI where the \code{\link{GtkLinkButton}} points.}
\usage{gtkLinkButtonSetUri(object, uri)}
\arguments{
\item{\code{object}}{[\code{\link{GtkLinkButton}}]  a \code{\link{GtkLinkButton}}}
\item{\code{uri}}{[character]  a valid URI}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
