\alias{gtkLinkButtonSetVisited}
\name{gtkLinkButtonSetVisited}
\title{gtkLinkButtonSetVisited}
\description{Sets the 'visited' state of the URI where the \code{\link{GtkLinkButton}}
points.  See \code{\link{gtkLinkButtonGetVisited}} for more details.}
\usage{gtkLinkButtonSetVisited(object, visited)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkLinkButton}}}
\item{\verb{visited}}{the new 'visited' state}
}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
