\alias{gtkListChildPosition}
\name{gtkListChildPosition}
\title{gtkListChildPosition}
\description{
Searches the children of \code{list} for the index of \code{child}.
\strong{WARNING: \code{gtk_list_child_position} is deprecated and should not be used in newly-written code.}
}
\usage{gtkListChildPosition(object, child)}
\arguments{
\item{\verb{object}}{the list widget.}
\item{\verb{child}}{the child to look for.}
}
\value{[integer] the index of the child, -1 if not found.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
