\alias{gtkListEndSelection}
\name{gtkListEndSelection}
\title{gtkListEndSelection}
\description{
Ends the selection. Used with \code{\link{gtkListExtendSelection}} and
\code{\link{gtkListStartSelection}}. Only in \verb{GTK_SELECTION_EXTENDED} mode.
\strong{WARNING: \code{gtk_list_end_selection} is deprecated and should not be used in newly-written code.}
}
\usage{gtkListEndSelection(object)}
\arguments{\item{\verb{object}}{the list widget.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
