\alias{gtkListItemDeselect}
\name{gtkListItemDeselect}
\title{gtkListItemDeselect}
\description{
Deselects the item, by emitting the item's "deselect" signal.
\strong{WARNING: \code{gtk_list_item_deselect} is deprecated and should not be used in newly-written code.}
}
\usage{gtkListItemDeselect(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkListItem}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
