\alias{gtkListNew}
\name{gtkListNew}
\title{gtkListNew}
\description{
Creates a new \verb{\link{GtkList}}.
\strong{WARNING: \code{gtk_list_new} is deprecated and should not be used in newly-written code.}
}
\usage{gtkListNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}] the newly-created \verb{\link{GtkList}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
