\alias{gtkListPrependItems}
\name{gtkListPrependItems}
\title{gtkListPrependItems}
\description{
Inserts \code{items} at the beginning of the \code{list}.
\strong{WARNING: \code{gtk_list_prepend_items} is deprecated and should not be used in newly-written code.  
Inserts \code{items} at the beginning of the \code{list}.}
}
\usage{gtkListPrependItems(object, items)}
\arguments{
\item{\code{object}}{[\code{\link{GtkList}}] the list widget.}
\item{\code{items}}{[list] the items.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
