\alias{gtkListSelectAll}
\name{gtkListSelectAll}
\title{gtkListSelectAll}
\description{
Selects all children of \code{list}. A signal will be emitted for each
newly selected child.
\strong{WARNING: \code{gtk_list_select_all} is deprecated and should not be used in newly-written code.}
}
\usage{gtkListSelectAll(object)}
\arguments{\item{\verb{object}}{the list widget.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
