\alias{gtkListStoreClear}
\name{gtkListStoreClear}
\title{gtkListStoreClear}
\description{Removes all rows from the list store.}
\usage{gtkListStoreClear(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkListStore}}]  a \code{\link{GtkListStore}}.}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
