\alias{gtkListStoreIterIsValid}
\name{gtkListStoreIterIsValid}
\title{gtkListStoreIterIsValid}
\description{\strong{WARNING: }}
\usage{gtkListStoreIterIsValid(object, iter)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkListStore}}]  A \verb{\link{GtkListStore}}.}
\item{\verb{iter}}{[\verb{\link{GtkTreeIter}}]  A \verb{\link{GtkTreeIter}}.}
}
\details{Checks if the given iter is a valid iter for this \verb{\link{GtkListStore}}.  

  Since  2.2}
\value{[logical] \code{TRUE} if the iter is valid, \code{FALSE} if the iter is invalid.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
