\alias{gtkListStoreNewv}
\name{gtkListStoreNewv}
\title{gtkListStoreNewv}
\description{Non-vararg creation function.  Used primarily by language bindings.}
\usage{gtkListStoreNewv(value)}
\arguments{\item{\verb{value}}{[\verb{\link{GType}}]  a list of \verb{\link{GType}} types for the columns, from first to last}}
\value{[\verb{\link{GtkListStore}}]  a new \verb{\link{GtkListStore}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
