\alias{gtkListToggleAddMode}
\name{gtkListToggleAddMode}
\title{gtkListToggleAddMode}
\description{
Toggles between adding to the selection and beginning a new selection. Only
in \verb{GTK_SELECTION_EXTENDED}. Useful with \code{\link{gtkListExtendSelection}}.
\strong{WARNING: \code{gtk_list_toggle_add_mode} is deprecated and should not be used in newly-written code.}
}
\usage{gtkListToggleAddMode(object)}
\arguments{\item{\verb{object}}{the list widget.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
