\alias{gtkListUnselectAll}
\name{gtkListUnselectAll}
\title{gtkListUnselectAll}
\description{
Unselects all children of \code{list}. A signal will be emitted for each
newly unselected child.
\strong{WARNING: \code{gtk_list_unselect_all} is deprecated and should not be used in newly-written code.}
}
\usage{gtkListUnselectAll(object)}
\arguments{\item{\verb{object}}{the list widget.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
