\alias{gtkMainLevel}
\name{gtkMainLevel}
\title{gtkMainLevel}
\description{Asks for the current nesting level of the main loop. This can be useful
when calling \code{\link{gtkQuitAdd}}.}
\usage{gtkMainLevel()}
\value{[numeric] the nesting level of the current invocation of the main loop.}
\note{
This function is deprecated. You should use \code{GDK_PRIORITY_REDRAW} instead.
This function is deprecated. You should use \code{G_PRIORITY_HIGH} instead.
This function is deprecated. You should use \code{G_PRIORITY_DEFAULT_IDLE} instead.
This function is deprecated. You should use \code{G_PRIORITY_LOW} instead.
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
