\alias{gtkMainQuit}
\name{gtkMainQuit}
\title{gtkMainQuit}
\description{Makes the innermost invocation of the main loop return when it regains 
control.}
\usage{gtkMainQuit()}
\note{
This function is deprecated. You should use \code{GDK_PRIORITY_REDRAW} instead.
This function is deprecated. You should use \code{G_PRIORITY_HIGH} instead.
This function is deprecated. You should use \code{G_PRIORITY_DEFAULT_IDLE} instead.
This function is deprecated. You should use \code{G_PRIORITY_LOW} instead.
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
