\alias{gtkMenuBarGetChildPackDirection}
\name{gtkMenuBarGetChildPackDirection}
\title{gtkMenuBarGetChildPackDirection}
\description{Retrieves the current child pack direction of the menubar.
See \code{\link{gtkMenuBarSetChildPackDirection}}.}
\usage{gtkMenuBarGetChildPackDirection(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMenuBar}}}}
\details{Since 2.8}
\value{[\code{\link{GtkPackDirection}}] the child pack direction}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
