\alias{gtkMenuBarGetPackDirection}
\name{gtkMenuBarGetPackDirection}
\title{gtkMenuBarGetPackDirection}
\description{Retrieves the current pack direction of the menubar. 
See \code{\link{gtkMenuBarSetPackDirection}}.}
\usage{gtkMenuBarGetPackDirection(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMenuBar}}}}
\details{Since 2.8}
\value{[\code{\link{GtkPackDirection}}] the pack direction}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
