\alias{gtkMenuBarNew}
\name{gtkMenuBarNew}
\title{gtkMenuBarNew}
\description{Creates the new \code{\link{GtkMenuBar}}}
\usage{gtkMenuBarNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] the \code{\link{GtkMenuBar}}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
