\alias{gtkMenuBarSetChildPackDirection}
\name{gtkMenuBarSetChildPackDirection}
\title{gtkMenuBarSetChildPackDirection}
\description{Sets how widgets should be packed inside the children of a menubar.}
\usage{gtkMenuBarSetChildPackDirection(object, child.pack.dir)}
\arguments{
\item{\code{object}}{[\code{\link{GtkMenuBar}}]  a \code{\link{GtkMenuBar}}.}
\item{\code{child.pack.dir}}{[\code{\link{GtkPackDirection}}]  a new \code{\link{GtkPackDirection}}.}
}
\details{  Since  2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
