\alias{gtkMenuDetach}
\name{gtkMenuDetach}
\title{gtkMenuDetach}
\description{Detaches the menu from the widget to which it had been attached.
This function will call the callback function, \code{detacher}, provided
when the \code{\link{gtkMenuAttachToWidget}} function was called.}
\usage{gtkMenuDetach(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMenu}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
