\alias{gtkMenuGetAccelGroup}
\name{gtkMenuGetAccelGroup}
\title{gtkMenuGetAccelGroup}
\description{Gets the \verb{\link{GtkAccelGroup}} which holds global accelerators for the menu.
See \code{\link{gtkMenuSetAccelGroup}}.}
\usage{gtkMenuGetAccelGroup(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkMenu}}] a \verb{\link{GtkMenu}}.}}
\value{[\verb{\link{GtkAccelGroup}}] the \verb{\link{GtkAccelGroup}} associated with the menu.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
