\alias{gtkMenuGetAccelPath}
\name{gtkMenuGetAccelPath}
\title{gtkMenuGetAccelPath}
\description{Retrieves the accelerator path set on the menu.}
\usage{gtkMenuGetAccelPath(object)}
\arguments{\item{\verb{object}}{a valid \code{\link{GtkMenu}}}}
\details{Since 2.14}
\value{[character] the accelerator path set on the menu.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
