\alias{gtkMenuGetAttachWidget}
\name{gtkMenuGetAttachWidget}
\title{gtkMenuGetAttachWidget}
\description{Returns the \verb{\link{GtkWidget}} that the menu is attached to.}
\usage{gtkMenuGetAttachWidget(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkMenu}}] a \verb{\link{GtkMenu}}.}}
\value{[\verb{\link{GtkWidget}}] the \verb{\link{GtkWidget}} that the menu is attached to.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
