\alias{gtkMenuGetMonitor}
\name{gtkMenuGetMonitor}
\title{gtkMenuGetMonitor}
\description{Retrieves the number of the monitor on which to show the menu.}
\usage{gtkMenuGetMonitor(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMenu}}}}
\details{Since 2.14}
\value{[integer] the number of the monitor on which the menu should
be popped up or -1, if no monitor has been set}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
