\alias{gtkMenuGetReserveToggleSize}
\name{gtkMenuGetReserveToggleSize}
\title{gtkMenuGetReserveToggleSize}
\description{Returns whether the menu reserves space for toggles and
icons, regardless of their actual presence.}
\usage{gtkMenuGetReserveToggleSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMenu}}}}
\details{Since 2.18}
\value{[logical] Whether the menu reserves toggle space}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
