\alias{gtkMenuGetTitle}
\name{gtkMenuGetTitle}
\title{gtkMenuGetTitle}
\description{Returns the title of the menu. See \code{\link{gtkMenuSetTitle}}.}
\usage{gtkMenuGetTitle(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMenu}}}}
\value{[character] the title of the menu, or \code{NULL} if the menu has no
title set on it.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
