\alias{gtkMenuItemDeselect}
\name{gtkMenuItemDeselect}
\title{gtkMenuItemDeselect}
\description{Emits the "deselect" signal on the given item. Behaves exactly like
\code{\link{gtkItemDeselect}}.}
\usage{gtkMenuItemDeselect(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkMenuItem}}] the menu item}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
