\alias{gtkMenuItemGetAccelPath}
\name{gtkMenuItemGetAccelPath}
\title{gtkMenuItemGetAccelPath}
\description{Retrieve the accelerator path that was previously set on \code{menu.item}.}
\usage{gtkMenuItemGetAccelPath(object)}
\arguments{\item{\verb{object}}{a valid \code{\link{GtkMenuItem}}}}
\details{See \code{\link{gtkMenuItemSetAccelPath}} for details.
  Since 2.14}
\value{[character] the accelerator path corresponding to this menu item's
functionality, or \code{NULL} if not set}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
