\alias{gtkMenuItemGetLabel}
\name{gtkMenuItemGetLabel}
\title{gtkMenuItemGetLabel}
\description{Sets \code{text} on the \code{menu.item} label}
\usage{gtkMenuItemGetLabel(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMenuItem}}}}
\details{Since 2.16}
\value{[character] The text in the \code{menu.item} label. This is the internal
string used by the label, and must not be modified.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
