\alias{gtkMenuItemGetRightJustified}
\name{gtkMenuItemGetRightJustified}
\title{gtkMenuItemGetRightJustified}
\description{Gets whether the menu item appears justified at the right
side of the menu bar.}
\usage{gtkMenuItemGetRightJustified(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMenuItem}}}}
\value{[logical] \code{TRUE} if the menu item will appear at the
far right if added to a menu bar.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
