\alias{gtkMenuItemGetUseUnderline}
\name{gtkMenuItemGetUseUnderline}
\title{gtkMenuItemGetUseUnderline}
\description{Checks if an underline in the text indicates the next character should be
used for the mnemonic accelerator key.}
\usage{gtkMenuItemGetUseUnderline(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMenuItem}}}}
\details{Since 2.16}
\value{[logical] \code{TRUE} if an embedded underline in the label indicates
the mnemonic accelerator key.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
