\alias{gtkMenuItemNew}
\name{gtkMenuItemNew}
\title{gtkMenuItemNew}
\description{Creates a new \verb{\link{GtkMenuItem}}.}
\usage{gtkMenuItemNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}] the newly created \verb{\link{GtkMenuItem}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
