\alias{gtkMenuItemNewWithLabel}
\name{gtkMenuItemNewWithLabel}
\title{gtkMenuItemNewWithLabel}
\description{Creates a new \verb{\link{GtkMenuItem}} whose child is a \verb{\link{GtkLabel}}.}
\usage{gtkMenuItemNewWithLabel(label, show = TRUE)}
\arguments{\item{\verb{label}}{[character] the text for the label}}
\value{[\verb{\link{GtkWidget}}] the newly created \verb{\link{GtkMenuItem}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
