\alias{gtkMenuItemRemoveSubmenu}
\name{gtkMenuItemRemoveSubmenu}
\title{gtkMenuItemRemoveSubmenu}
\description{Removes the widget's submenu.}
\usage{gtkMenuItemRemoveSubmenu(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkMenuItem}}] the menu item widget}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
