\alias{gtkMenuItemSelect}
\name{gtkMenuItemSelect}
\title{gtkMenuItemSelect}
\description{Emits the "select" signal on the given item. Behaves exactly like
\code{\link{gtkItemSelect}}.}
\usage{gtkMenuItemSelect(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkMenuItem}}] the menu item}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
