\alias{gtkMenuItemSetLabel}
\name{gtkMenuItemSetLabel}
\title{gtkMenuItemSetLabel}
\description{Sets \code{text} on the \code{menu.item} label}
\usage{gtkMenuItemSetLabel(object, label)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkMenuItem}}}
\item{\verb{label}}{the text you want to set}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
