\alias{gtkMenuItemSetSubmenu}
\name{gtkMenuItemSetSubmenu}
\title{gtkMenuItemSetSubmenu}
\description{Sets the widget submenu, or changes it.}
\usage{gtkMenuItemSetSubmenu(object, submenu)}
\arguments{
\item{\code{object}}{[\code{\link{GtkMenuItem}}] the menu item widget}
\item{\code{submenu}}{[\code{\link{GtkWidget}}] the submenu}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
