\alias{gtkMenuItemSetUseUnderline}
\name{gtkMenuItemSetUseUnderline}
\title{gtkMenuItemSetUseUnderline}
\description{If true, an underline in the text indicates the next character should be
used for the mnemonic accelerator key.}
\usage{gtkMenuItemSetUseUnderline(object, setting)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkMenuItem}}}
\item{\verb{setting}}{\code{TRUE} if underlines in the text indicate mnemonics}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
