\alias{gtkMenuItemToggleSizeAllocate}
\name{gtkMenuItemToggleSizeAllocate}
\title{gtkMenuItemToggleSizeAllocate}
\description{Emits the "toggle\_size\_allocate" signal on the given item.}
\usage{gtkMenuItemToggleSizeAllocate(object, allocation)}
\arguments{
\item{\code{object}}{[\code{\link{GtkMenuItem}}] the menu item.}
\item{\code{allocation}}{[integer] the allocation to use as signal data.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
