\alias{gtkMenuPopdown}
\name{gtkMenuPopdown}
\title{gtkMenuPopdown}
\description{Removes the menu from the screen.}
\usage{gtkMenuPopdown(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkMenu}}] a \code{\link{GtkMenu}}.}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
