\alias{gtkMenuReorderChild}
\name{gtkMenuReorderChild}
\title{gtkMenuReorderChild}
\description{Moves a \verb{\link{GtkMenuItem}} to a new position within the \verb{\link{GtkMenu}}.}
\usage{gtkMenuReorderChild(object, child, position)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkMenu}}] a \verb{\link{GtkMenu}}.}
\item{\verb{child}}{[\verb{\link{GtkWidget}}] the \verb{\link{GtkMenuItem}} to move.}
\item{\verb{position}}{[integer] the new position to place \code{child}.  Positions are numbered from 
0 to n-1.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
