\alias{gtkMenuReposition}
\name{gtkMenuReposition}
\title{gtkMenuReposition}
\description{Repositions the menu according to its position function.}
\usage{gtkMenuReposition(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkMenu}}] a \code{\link{GtkMenu}}.}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
