\alias{gtkMenuSetActive}
\name{gtkMenuSetActive}
\title{gtkMenuSetActive}
\description{Selects the specified menu item within the menu.  This is used by the
\verb{\link{GtkOptionMenu}} and should not be used by anyone else.}
\usage{gtkMenuSetActive(object, index)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkMenu}}] a \verb{\link{GtkMenu}}.}
\item{\verb{index}}{[numeric] the index of the menu item to select.  Index values are from
0 to n-1.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
