\alias{gtkMenuSetReserveToggleSize}
\name{gtkMenuSetReserveToggleSize}
\title{gtkMenuSetReserveToggleSize}
\description{Sets whether the menu should reserve space for drawing toggles 
or icons, regardless of their actual presence.}
\usage{gtkMenuSetReserveToggleSize(object, reserve.toggle.size)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkMenu}}}
\item{\verb{reserve.toggle.size}}{whether to reserve size for toggles}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
