\alias{gtkMenuSetScreen}
\name{gtkMenuSetScreen}
\title{gtkMenuSetScreen}
\description{Sets the \verb{\link{GdkScreen}} on which the menu will be displayed.}
\usage{gtkMenuSetScreen(object, screen = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkMenu}}]  a \verb{\link{GtkMenu}}.}
\item{\verb{screen}}{[\verb{\link{GdkScreen}}]  a \verb{\link{GdkScreen}}, or \code{NULL} if the screen should be
         determined by the widget the menu is attached to.}
}
\details{  Since  2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
