\alias{gtkMenuSetTearoffState}
\name{gtkMenuSetTearoffState}
\title{gtkMenuSetTearoffState}
\description{Changes the tearoff state of the menu.  A menu is normally displayed 
as drop down menu which persists as long as the menu is active.  It can 
also be displayed as a tearoff menu which persists until it is closed 
or reattached.}
\usage{gtkMenuSetTearoffState(object, torn.off)}
\arguments{
\item{\code{object}}{[\code{\link{GtkMenu}}] a \code{\link{GtkMenu}}.}
\item{\code{torn.off}}{[logical] If \code{TRUE}, menu is displayed as a tearoff menu.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
