\alias{gtkMenuSetTitle}
\name{gtkMenuSetTitle}
\title{gtkMenuSetTitle}
\description{Sets the title string for the menu.  The title is displayed when the menu
is shown as a tearoff menu.}
\usage{gtkMenuSetTitle(object, title)}
\arguments{
\item{\code{object}}{[\code{\link{GtkMenu}}]  a \code{\link{GtkMenu}}}
\item{\code{title}}{[character]  a string containing the title for the menu.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
