\alias{gtkMenuShellActivateItem}
\name{gtkMenuShellActivateItem}
\title{gtkMenuShellActivateItem}
\description{Activates the menu item within the menu shell.}
\usage{gtkMenuShellActivateItem(object, menu.item, force.deactivate)}
\arguments{
\item{\code{object}}{[\code{\link{GtkMenuShell}}] a \code{\link{GtkMenuShell}}.}
\item{\code{menu.item}}{[\code{\link{GtkWidget}}] The \code{\link{GtkMenuItem}} to activate.}
\item{\code{force.deactivate}}{[logical] If TRUE, force the deactivation of the menu shell
after the menu item is activated.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
