\alias{gtkMenuShellCancel}
\name{gtkMenuShellCancel}
\title{gtkMenuShellCancel}
\description{Cancels the selection within the menu shell.}
\usage{gtkMenuShellCancel(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkMenuShell}}]  a \code{\link{GtkMenuShell}}}}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
