\alias{gtkMenuShellDeselect}
\name{gtkMenuShellDeselect}
\title{gtkMenuShellDeselect}
\description{Deselects the currently selected item from the menu shell, if any.}
\usage{gtkMenuShellDeselect(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMenuShell}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
