\alias{gtkMenuShellInsert}
\name{gtkMenuShellInsert}
\title{gtkMenuShellInsert}
\description{Adds a new \verb{\link{GtkMenuItem}} to the menu shell's item list at the position
indicated by \code{position}. }
\usage{gtkMenuShellInsert(object, child, position)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkMenuShell}}] a \verb{\link{GtkMenuShell}}.}
\item{\verb{child}}{[\verb{\link{GtkWidget}}] The \verb{\link{GtkMenuItem}} to add.}
\item{\verb{position}}{[integer] The position in the item list where \code{child} is added.
Positions are numbered from 0 to n-1.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
