\alias{gtkMenuShellPrepend}
\name{gtkMenuShellPrepend}
\title{gtkMenuShellPrepend}
\description{Adds a new \verb{\link{GtkMenuItem}} to the beginning of the menu shell's item list.}
\usage{gtkMenuShellPrepend(object, child)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkMenuShell}}] a \verb{\link{GtkMenuShell}}.}
\item{\verb{child}}{[\verb{\link{GtkWidget}}] The \verb{\link{GtkMenuItem}} to add.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
