\alias{gtkMenuShellSelectItem}
\name{gtkMenuShellSelectItem}
\title{gtkMenuShellSelectItem}
\description{Selects the menu item from the menu shell.}
\usage{gtkMenuShellSelectItem(object, menu.item)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkMenuShell}}] a \verb{\link{GtkMenuShell}}.}
\item{\verb{menu.item}}{[\verb{\link{GtkWidget}}] The \verb{\link{GtkMenuItem}} to select.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
