\alias{gtkMenuToolButtonGetMenu}
\name{gtkMenuToolButtonGetMenu}
\title{gtkMenuToolButtonGetMenu}
\description{Gets the \verb{\link{GtkMenu}} associated with \verb{\link{GtkMenuToolButton}}.}
\usage{gtkMenuToolButtonGetMenu(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkMenuToolButton}}]  a \verb{\link{GtkMenuToolButton}}}}
\details{  Since  2.6}
\value{[\verb{\link{GtkWidget}}]  the \verb{\link{GtkMenu}} associated with \verb{\link{GtkMenuToolButton}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
