\alias{gtkMenuToolButtonNew}
\name{gtkMenuToolButtonNew}
\title{gtkMenuToolButtonNew}
\description{Creates a new \code{\link{GtkMenuToolButton}} using \code{icon.widget} as icon and
\code{label} as label.}
\usage{gtkMenuToolButtonNew(icon.widget, label, show = TRUE)}
\arguments{
\item{\code{icon.widget}}{[\code{\link{GtkWidget}}]  a widget that will be used as icon widget, or \code{NULL}}
\item{\code{label}}{[character]  a string that will be used as label, or \code{NULL}}
}
\details{  Since  2.6}
\value{[\code{\link{GtkToolItem}}]  the new \code{\link{GtkMenuToolButton}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
