\alias{gtkMenuToolButtonNewFromStock}
\name{gtkMenuToolButtonNewFromStock}
\title{gtkMenuToolButtonNewFromStock}
\description{Creates a new \verb{\link{GtkMenuToolButton}}.
The new \verb{\link{GtkMenuToolButton}} will contain an icon and label from
the stock item indicated by \code{stock.id}.}
\usage{gtkMenuToolButtonNewFromStock(stock.id)}
\arguments{\item{\verb{stock.id}}{[character]  the name of a stock item}}
\details{  Since  2.6}
\value{[\verb{\link{GtkToolItem}}]  the new \verb{\link{GtkMenuToolButton}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
