\alias{gtkMenuToolButtonSetArrowTooltipMarkup}
\name{gtkMenuToolButtonSetArrowTooltipMarkup}
\title{gtkMenuToolButtonSetArrowTooltipMarkup}
\description{Sets the tooltip markup text to be used as tooltip for the arrow button
which pops up the menu.  See \code{\link{gtkToolItemSetTooltip}} for setting a
tooltip on the whole \verb{\link{GtkMenuToolButton}}.}
\usage{gtkMenuToolButtonSetArrowTooltipMarkup(object, markup)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkMenuToolButton}}]  a \verb{\link{GtkMenuToolButton}}}
\item{\verb{markup}}{[character]  markup text to be used as tooltip text for button's arrow button}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
