\alias{gtkMenuToolButtonSetArrowTooltipText}
\name{gtkMenuToolButtonSetArrowTooltipText}
\title{gtkMenuToolButtonSetArrowTooltipText}
\description{Sets the tooltip text to be used as tooltip for the arrow button which
pops up the menu.  See \code{\link{gtkToolItemSetTooltip}} for setting a tooltip
on the whole \code{\link{GtkMenuToolButton}}.}
\usage{gtkMenuToolButtonSetArrowTooltipText(object, text)}
\arguments{
\item{\code{object}}{[\code{\link{GtkMenuToolButton}}]  a \code{\link{GtkMenuToolButton}}}
\item{\code{text}}{[character]  text to be used as tooltip text for button's arrow button}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
