\alias{gtkMenuToolButtonSetMenu}
\name{gtkMenuToolButtonSetMenu}
\title{gtkMenuToolButtonSetMenu}
\description{Sets the \verb{\link{GtkMenu}} that is popped up when the user clicks on the arrow.
If \code{menu} is NULL, the arrow button becomes insensitive.}
\usage{gtkMenuToolButtonSetMenu(object, menu)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkMenuToolButton}}]  a \verb{\link{GtkMenuToolButton}}}
\item{\verb{menu}}{[\verb{\link{GtkWidget}}]  the \verb{\link{GtkMenu}} associated with \verb{\link{GtkMenuToolButton}}}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
