\alias{gtkMessageDialogGetImage}
\name{gtkMessageDialogGetImage}
\title{gtkMessageDialogGetImage}
\description{Gets the dialog's image.}
\usage{gtkMessageDialogGetImage(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMessageDialog}}}}
\details{Since 2.14}
\value{[\code{\link{GtkWidget}}] the dialog's image}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
