\alias{gtkMessageDialogSetImage}
\name{gtkMessageDialogSetImage}
\title{gtkMessageDialogSetImage}
\description{Sets the dialog's image to \code{image}.}
\usage{gtkMessageDialogSetImage(object, image)}
\arguments{
\item{\code{object}}{[\code{\link{GtkMessageDialog}}]  a \code{\link{GtkMessageDialog}}}
\item{\code{image}}{[\code{\link{GtkWidget}}]  the image}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
