\alias{gtkMessageDialogSetMarkup}
\name{gtkMessageDialogSetMarkup}
\title{gtkMessageDialogSetMarkup}
\description{Sets the text of the message dialog to be \code{str}, which is marked
up with the Pango text markup
language.}
\usage{gtkMessageDialogSetMarkup(object, str)}
\arguments{
\item{\code{object}}{[\code{\link{GtkMessageDialog}}]  a \code{\link{GtkMessageDialog}}}
\item{\code{str}}{[character]  markup string (see Pango markup format)}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
