\alias{gtkMiscGetAlignment}
\name{gtkMiscGetAlignment}
\title{gtkMiscGetAlignment}
\description{Gets the X and Y alignment of the widget within its allocation. See
\code{\link{gtkMiscSetAlignment}}.}
\usage{gtkMiscGetAlignment(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkMisc}}]  a \code{\link{GtkMisc}}}}
\value{
A list containing the following elements:
\item{\code{xalign}}{[numeric]  location to store X alignment of \code{misc}, or \code{NULL}}
\item{\code{yalign}}{[numeric]  location to store Y alignment of \code{misc}, or \code{NULL}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
