\alias{gtkMiscGetPadding}
\name{gtkMiscGetPadding}
\title{gtkMiscGetPadding}
\description{Gets the padding in the X and Y directions of the widget. See \code{\link{gtkMiscSetPadding}}.}
\usage{gtkMiscGetPadding(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkMisc}}]  a \code{\link{GtkMisc}}}}
\value{
A list containing the following elements:
\item{\code{xpad}}{[integer]  location to store padding in the X direction, or \code{NULL}}
\item{\code{ypad}}{[integer]  location to store padding in the Y direction, or \code{NULL}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
