\alias{gtkMiscSetAlignment}
\name{gtkMiscSetAlignment}
\title{gtkMiscSetAlignment}
\description{Sets the alignment of the widget.}
\usage{gtkMiscSetAlignment(object, xalign, yalign)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkMisc}}.}
\item{\verb{xalign}}{the horizontal alignment, from 0 (left) to 1 (right).}
\item{\verb{yalign}}{the vertical alignment, from 0 (top) to 1 (bottom).}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
