\alias{gtkMiscSetPadding}
\name{gtkMiscSetPadding}
\title{gtkMiscSetPadding}
\description{Sets the amount of space to add around the widget.}
\usage{gtkMiscSetPadding(object, xpad, ypad)}
\arguments{
\item{\code{object}}{[\code{\link{GtkMisc}}] a \code{\link{GtkMisc}}.}
\item{\code{xpad}}{[integer] the amount of space to add on the left and right of the widget,
in pixels.}
\item{\code{ypad}}{[integer] the amount of space to add on the top and bottom of the widget,
in pixels.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
