\alias{gtkMountOperationGetParent}
\name{gtkMountOperationGetParent}
\title{gtkMountOperationGetParent}
\description{Gets the transient parent used by the \code{\link{GtkMountOperation}}}
\usage{gtkMountOperationGetParent(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMountOperation}}}}
\details{Since 2.14}
\value{[\code{\link{GtkWindow}}] the transient parent for windows shown by \code{op}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
