\alias{gtkMountOperationGetScreen}
\name{gtkMountOperationGetScreen}
\title{gtkMountOperationGetScreen}
\description{Gets the screen on which windows of the \code{\link{GtkMountOperation}} 
will be shown.}
\usage{gtkMountOperationGetScreen(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMountOperation}}}}
\details{Since 2.14}
\value{[\code{\link{GdkScreen}}] the screen on which windows of \code{op} are shown}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
