\alias{gtkMountOperationIsShowing}
\name{gtkMountOperationIsShowing}
\title{gtkMountOperationIsShowing}
\description{Returns whether the \code{\link{GtkMountOperation}} is currently displaying
a window.}
\usage{gtkMountOperationIsShowing(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMountOperation}}}}
\details{Since 2.14}
\value{[logical] \code{TRUE} if \code{op} is currently displaying a window}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
