\alias{gtkMountOperationNew}
\name{gtkMountOperationNew}
\title{gtkMountOperationNew}
\description{Creates a new \code{\link{GtkMountOperation}}}
\usage{gtkMountOperationNew(parent = NULL)}
\arguments{\item{\verb{parent}}{transient parent of the window, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}}
\details{Since 2.14}
\value{[\code{\link{GMountOperation}}] a new \code{\link{GtkMountOperation}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
