\alias{gtkMountOperationSetParent}
\name{gtkMountOperationSetParent}
\title{gtkMountOperationSetParent}
\description{Sets the transient parent for windows shown by the
\code{\link{GtkMountOperation}}.}
\usage{gtkMountOperationSetParent(object, parent)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkMountOperation}}}
\item{\verb{parent}}{transient parent of the window, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
