\alias{gtkMountOperationSetScreen}
\name{gtkMountOperationSetScreen}
\title{gtkMountOperationSetScreen}
\description{Sets the screen to show windows of the \code{\link{GtkMountOperation}} on.}
\usage{gtkMountOperationSetScreen(object, screen)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkMountOperation}}}
\item{\verb{screen}}{a \code{\link{GdkScreen}}}
}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
