\name{GtkMozEmbed}
\title{The Embeddable GTK Mozilla Widget}
\alias{GtkMozEmbedChromeFlags}
\alias{GtkMozEmbedProgressFlags}
\alias{GtkMozEmbedReloadFlags}
\alias{GtkMozEmbedStatusFlags}
\alias{gtkMozEmbed}
\alias{gtkMozEmbedAppendData}
\alias{gtkMozEmbedCanGoBack}
\alias{gtkMozEmbedCanGoForward}
\alias{gtkMozEmbedCloseStream}
\alias{gtkMozEmbedGetChromeMask}
\alias{gtkMozEmbedGetJsStatus}
\alias{gtkMozEmbedGetLinkMessage}
\alias{gtkMozEmbedGetLocation}
\alias{gtkMozEmbedGetTitle}
\alias{gtkMozEmbedGetType}
\alias{gtkMozEmbedGoBackgtkMozEmbedGoForward}
\alias{gtkMozEmbedLoadUrl}
\alias{gtkMozEmbedNew}
\alias{gtkMozEmbedOpenStream}
\alias{gtkMozEmbedPopStartup}
\alias{gtkMozEmbedPushStartup}
\alias{gtkMozEmbedReload}
\alias{gtkMozEmbedRenderData}
\alias{gtkMozEmbedSetChromeMask}
\alias{gtkMozEmbedSetCompPath}
\alias{gtkMozEmbedSetProfilePath}
\alias{gtkMozEmbedSingleGet}
\alias{gtkMozEmbedStopLoad}
\alias{gtkMozEmbedGoBack} 
\alias{gtkMozEmbedGoForward} 
\alias{gtkNotebookCurrentPage}
\description{Please see \url{http://www.mozilla.org/unix/gtk-embedding.html} for GtkMozEmbed documentation}
\author{Michael Lawrence}
\keyword{internal}
