\alias{gtkNotebookGetGroup}
\name{gtkNotebookGetGroup}
\title{gtkNotebookGetGroup}
\description{Gets the current group identificator pointer for \code{notebook}.}
\usage{gtkNotebookGetGroup(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkNotebook}}]  a \code{\link{GtkNotebook}}}}
\details{  Since  2.12}
\value{[R object]  the group identificator, or \code{NULL} if none is set.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
