\alias{gtkNotebookGetGroupId}
\name{gtkNotebookGetGroupId}
\title{gtkNotebookGetGroupId}
\description{Gets the current group identificator for \code{notebook}.}
\usage{gtkNotebookGetGroupId(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkNotebook}}]  a \code{\link{GtkNotebook}}}}
\details{  Since  2.10}
\value{[integer]  the group identificator, or -1 if none is set.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
