\alias{gtkNotebookGetNPages}
\name{gtkNotebookGetNPages}
\title{gtkNotebookGetNPages}
\description{Gets the number of pages in a notebook.}
\usage{gtkNotebookGetNPages(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkNotebook}}]  a \code{\link{GtkNotebook}}}}
\details{  Since  2.2}
\value{[integer]  the number of pages in the notebook.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
