\alias{gtkNotebookGetTabDetachable}
\name{gtkNotebookGetTabDetachable}
\title{gtkNotebookGetTabDetachable}
\description{Returns whether the tab contents can be detached from \code{notebook}.}
\usage{gtkNotebookGetTabDetachable(object, child)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkNotebook}}]  a \verb{\link{GtkNotebook}}}
\item{\verb{child}}{[\verb{\link{GtkWidget}}]  a child \verb{\link{GtkWidget}}}
}
\details{  Since  2.10}
\value{[logical]  TRUE if the tab is detachable.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
