\alias{gtkNotebookGetTabPos}
\name{gtkNotebookGetTabPos}
\title{gtkNotebookGetTabPos}
\description{Gets the edge at which the tabs for switching pages in the
notebook are drawn.}
\usage{gtkNotebookGetTabPos(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkNotebook}}]  a \code{\link{GtkNotebook}}}}
\value{[\code{\link{GtkPositionType}}]  the edge at which the tabs are drawn}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
