\alias{gtkNotebookNew}
\name{gtkNotebookNew}
\title{gtkNotebookNew}
\description{Creates a new \verb{\link{GtkNotebook}} widget with no pages.}
\usage{gtkNotebookNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}]  the newly created \verb{\link{GtkNotebook}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
