\alias{gtkNotebookPopupDisable}
\name{gtkNotebookPopupDisable}
\title{gtkNotebookPopupDisable}
\description{Disables the popup menu.}
\usage{gtkNotebookPopupDisable(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkNotebook}}]  a \code{\link{GtkNotebook}}}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
