\alias{gtkNotebookPopupEnable}
\name{gtkNotebookPopupEnable}
\title{gtkNotebookPopupEnable}
\description{Enables the popup menu: if the user clicks with the right mouse button on
the bookmarks, a menu with all the pages will be popped up.}
\usage{gtkNotebookPopupEnable(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkNotebook}}]  a \code{\link{GtkNotebook}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
