\alias{gtkNotebookRemovePage}
\name{gtkNotebookRemovePage}
\title{gtkNotebookRemovePage}
\description{Removes a page from the notebook given its index
in the notebook.}
\usage{gtkNotebookRemovePage(object, page.num)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkNotebook}}]  a \verb{\link{GtkNotebook}}.}
\item{\verb{page.num}}{[integer]  the index of a notebook page, starting
           from 0. If -1, the last page will
           be removed.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
