\alias{gtkNotebookSetHomogeneousTabs}
\name{gtkNotebookSetHomogeneousTabs}
\title{gtkNotebookSetHomogeneousTabs}
\description{
Sets whether the tabs must have all the same size or not.
\strong{WARNING: \code{gtk_notebook_set_homogeneous_tabs} is deprecated and should not be used in newly-written code.}
}
\usage{gtkNotebookSetHomogeneousTabs(object, homogeneous)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkNotebook}}}
\item{\verb{homogeneous}}{\code{TRUE} if all tabs should be the same size.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
