\alias{gtkNotebookSetMenuLabel}
\name{gtkNotebookSetMenuLabel}
\title{gtkNotebookSetMenuLabel}
\description{Changes the menu label for the page containing \code{child}.}
\usage{gtkNotebookSetMenuLabel(object, child, menu.label = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkNotebook}}]  a \code{\link{GtkNotebook}}}
\item{\code{child}}{[\code{\link{GtkWidget}}]  the child widget}
\item{\code{menu.label}}{[\code{\link{GtkWidget}}]  the menu label, or NULL for default}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
