\alias{gtkNotebookSetMenuLabelText}
\name{gtkNotebookSetMenuLabelText}
\title{gtkNotebookSetMenuLabelText}
\description{Creates a new label and sets it as the menu label of \code{child}.}
\usage{gtkNotebookSetMenuLabelText(object, child, menu.text)}
\arguments{
\item{\code{object}}{[\code{\link{GtkNotebook}}]  a \code{\link{GtkNotebook}}}
\item{\code{child}}{[\code{\link{GtkWidget}}]  the child widget}
\item{\code{menu.text}}{[character]  the label text}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
