\alias{gtkNotebookSetScrollable}
\name{gtkNotebookSetScrollable}
\title{gtkNotebookSetScrollable}
\description{Sets whether the tab label area will have arrows for scrolling if
there are too many tabs to fit in the area.}
\usage{gtkNotebookSetScrollable(object, scrollable)}
\arguments{
\item{\code{object}}{[\code{\link{GtkNotebook}}]  a \code{\link{GtkNotebook}}}
\item{\code{scrollable}}{[logical] \code{TRUE} if scroll arrows should be added}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
