\alias{gtkNotebookSetShowBorder}
\name{gtkNotebookSetShowBorder}
\title{gtkNotebookSetShowBorder}
\description{Sets whether a bevel will be drawn around the notebook pages.
This only has a visual effect when the tabs are not shown.
See \code{\link{gtkNotebookSetShowTabs}}.}
\usage{gtkNotebookSetShowBorder(object, show.border)}
\arguments{
\item{\code{object}}{[\code{\link{GtkNotebook}}]  a \code{\link{GtkNotebook}}}
\item{\code{show.border}}{[logical]  \code{TRUE} if a bevel should be drawn around the notebook.}
}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
