\alias{gtkNotebookSetTabHborder}
\name{gtkNotebookSetTabHborder}
\title{gtkNotebookSetTabHborder}
\description{
Sets the width of the horizontal border of tab labels.
\strong{WARNING: \code{gtk_notebook_set_tab_hborder} is deprecated and should not be used in newly-written code.}
}
\usage{gtkNotebookSetTabHborder(object, tab.hborder)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkNotebook}}]  a \verb{\link{GtkNotebook}}}
\item{\verb{tab.hborder}}{[numeric]  width of the horizontal border of tab labels.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
