\alias{gtkNotebookSetTabLabelPacking}
\name{gtkNotebookSetTabLabelPacking}
\title{gtkNotebookSetTabLabelPacking}
\description{Sets the packing parameters for the tab label of the page
containing \code{child}. See \code{\link{gtkBoxPackStart}} for the exact meaning
of the parameters.}
\usage{gtkNotebookSetTabLabelPacking(object, child, expand, fill, pack.type)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkNotebook}}]  a \verb{\link{GtkNotebook}}}
\item{\verb{child}}{[\verb{\link{GtkWidget}}]  the child widget}
\item{\verb{expand}}{[logical]  whether to expand the bookmark or not}
\item{\verb{fill}}{[logical]  whether the bookmark should fill the allocated area or not}
\item{\verb{pack.type}}{[\code{\link{GtkPackType}}]  the position of the bookmark}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
