\alias{gtkNotebookSetTabLabelText}
\name{gtkNotebookSetTabLabelText}
\title{gtkNotebookSetTabLabelText}
\description{Creates a new label and sets it as the tab label for the page
containing \code{child}.}
\usage{gtkNotebookSetTabLabelText(object, child, tab.text)}
\arguments{
\item{\code{object}}{[\code{\link{GtkNotebook}}]  a \code{\link{GtkNotebook}}}
\item{\code{child}}{[\code{\link{GtkWidget}}]  the page}
\item{\code{tab.text}}{[character]  the label text}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
