\alias{gtkNotebookSetTabPos}
\name{gtkNotebookSetTabPos}
\title{gtkNotebookSetTabPos}
\description{Sets the edge at which the tabs for switching pages in the
notebook are drawn.}
\usage{gtkNotebookSetTabPos(object, pos)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkNotebook}}]  a \verb{\link{GtkNotebook}}.}
\item{\verb{pos}}{[\code{\link{GtkPositionType}}]  the edge to draw the tabs at.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
