\alias{gtkNotebookSetTabReorderable}
\name{gtkNotebookSetTabReorderable}
\title{gtkNotebookSetTabReorderable}
\description{Sets whether the notebook tab can be reordered
via drag and drop or not.}
\usage{gtkNotebookSetTabReorderable(object, child, reorderable)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkNotebook}}]  a \verb{\link{GtkNotebook}}}
\item{\verb{child}}{[\verb{\link{GtkWidget}}]  a child \verb{\link{GtkWidget}}}
\item{\verb{reorderable}}{[logical]  whether the tab is reorderable or not.}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
