\alias{gtkNotebookSetTabVborder}
\name{gtkNotebookSetTabVborder}
\title{gtkNotebookSetTabVborder}
\description{
Sets the width of the vertical border of tab labels.
\strong{WARNING: \code{gtk_notebook_set_tab_vborder} is deprecated and should not be used in newly-written code.}
}
\usage{gtkNotebookSetTabVborder(object, tab.vborder)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkNotebook}}}
\item{\verb{tab.vborder}}{width of the vertical border of tab labels.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
