\name{GtkObject}
\alias{GtkObject}
\title{GtkObject}
\description{
As far as the RGtk2 user is concerned, the \code{GtkObject} class just indicates 
that an object is part of the \code{\link{GTK}} library, though the library 
also contains non-\code{GtkObject}s.}
\author{Michael Lawrence}
\keyword{interface}
\keyword{internal}
