\alias{gtkOffscreenWindowNew}
\name{gtkOffscreenWindowNew}
\title{gtkOffscreenWindowNew}
\description{Creates a toplevel container widget that is used to retrieve
snapshots of widgets without showing them on the screen.  For
widgets that are on the screen and part of a normal widget
hierarchy, \code{\link{gtkWidgetGetSnapshot}} can be used instead.}
\usage{gtkOffscreenWindowNew(show = TRUE)}
\details{Since 2.20}
\value{[\code{\link{GtkWidget}}] A pointer to a \code{\link{GtkWidget}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
