\alias{gtkOldEditableChanged}
\name{gtkOldEditableChanged}
\title{gtkOldEditableChanged}
\description{
Emits the ::changed signal on \code{old.editable}.
\strong{WARNING: \code{gtk_old_editable_changed} is deprecated and should not be used in newly-written code.}
}
\usage{gtkOldEditableChanged(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkOldEditable}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
