\alias{gtkOptionMenuGetMenu}
\name{gtkOptionMenuGetMenu}
\title{gtkOptionMenuGetMenu}
\description{
Returns the \verb{\link{GtkMenu}} associated with the \verb{\link{GtkOptionMenu}}.
\strong{WARNING: \code{gtk_option_menu_get_menu} has been deprecated since version 2.4 and should not be used in newly-written code. Use \verb{\link{GtkComboBox}} instead.}
}
\usage{gtkOptionMenuGetMenu(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkOptionMenu}}] a \verb{\link{GtkOptionMenu}}.}}
\value{[\verb{\link{GtkWidget}}] the \verb{\link{GtkMenu}} associated with the \verb{\link{GtkOptionMenu}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
