\alias{gtkOptionMenuNew}
\name{gtkOptionMenuNew}
\title{gtkOptionMenuNew}
\description{
Creates a new \verb{\link{GtkOptionMenu}}.
\strong{WARNING: \code{gtk_option_menu_new} has been deprecated since version 2.4 and should not be used in newly-written code. Use \verb{\link{GtkComboBox}} instead.}
}
\usage{gtkOptionMenuNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}] a new \verb{\link{GtkOptionMenu}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
