\alias{gtkOptionMenuRemoveMenu}
\name{gtkOptionMenuRemoveMenu}
\title{gtkOptionMenuRemoveMenu}
\description{
Removes the menu from the option menu.
\strong{WARNING: \code{gtk_option_menu_remove_menu} is deprecated and should not be used in newly-written code. Use \code{\link{GtkComboBox}} instead.}
}
\usage{gtkOptionMenuRemoveMenu(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkOptionMenu}}] a \code{\link{GtkOptionMenu}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
