\alias{gtkOrientableGetOrientation}
\name{gtkOrientableGetOrientation}
\title{gtkOrientableGetOrientation}
\description{Retrieves the orientation of the \code{orientable}.}
\usage{gtkOrientableGetOrientation(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkOrientable}}}}
\details{Since 2.16}
\value{[\code{\link{GtkOrientation}}] the orientation of the \code{orientable}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
