\alias{gtkOrientableSetOrientation}
\name{gtkOrientableSetOrientation}
\title{gtkOrientableSetOrientation}
\description{Sets the orientation of the \code{orientable}.}
\usage{gtkOrientableSetOrientation(object, orientation)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkOrientable}}}
\item{\verb{orientation}}{the orientable's new orientation.}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
