\alias{gtkPageSetupCopy}
\name{gtkPageSetupCopy}
\title{gtkPageSetupCopy}
\description{Copies a \verb{\link{GtkPageSetup}}.}
\usage{gtkPageSetupCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkPageSetup}}]  the \verb{\link{GtkPageSetup}} to copy}}
\details{  Since  2.10}
\value{[\verb{\link{GtkPageSetup}}]  a copy of \code{other}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
