\alias{gtkPageSetupGetBottomMargin}
\name{gtkPageSetupGetBottomMargin}
\title{gtkPageSetupGetBottomMargin}
\description{Gets the bottom margin in units of \code{unit}.}
\usage{gtkPageSetupGetBottomMargin(object, unit)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPageSetup}}]  a \verb{\link{GtkPageSetup}}}
\item{\verb{unit}}{[\code{\link{GtkUnit}}]  the unit for the return value}
}
\details{  Since  2.10}
\value{[numeric]  the bottom margin}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
