\alias{gtkPageSetupGetOrientation}
\name{gtkPageSetupGetOrientation}
\title{gtkPageSetupGetOrientation}
\description{Gets the page orientation of the \code{\link{GtkPageSetup}}.}
\usage{gtkPageSetupGetOrientation(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkPageSetup}}]  a \code{\link{GtkPageSetup}}}}
\details{  Since  2.10}
\value{[\code{\link{GtkPageOrientation}}]  the page orientation}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
