\alias{gtkPageSetupGetRightMargin}
\name{gtkPageSetupGetRightMargin}
\title{gtkPageSetupGetRightMargin}
\description{Gets the right margin in units of \code{unit}.}
\usage{gtkPageSetupGetRightMargin(object, unit)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPageSetup}}]  a \code{\link{GtkPageSetup}}}
\item{\code{unit}}{[\code{\link{GtkUnit}}]  the unit for the return value}
}
\details{  Since  2.10}
\value{[numeric]  the right margin}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
