\alias{gtkPageSetupGetTopMargin}
\name{gtkPageSetupGetTopMargin}
\title{gtkPageSetupGetTopMargin}
\description{Gets the top margin in units of \code{unit}.}
\usage{gtkPageSetupGetTopMargin(object, unit)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPageSetup}}]  a \code{\link{GtkPageSetup}}}
\item{\code{unit}}{[\code{\link{GtkUnit}}]  the unit for the return value}
}
\details{  Since  2.10}
\value{[numeric]  the top margin}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
