\alias{gtkPageSetupNew}
\name{gtkPageSetupNew}
\title{gtkPageSetupNew}
\description{Creates a new \code{\link{GtkPageSetup}}.}
\usage{gtkPageSetupNew()}
\details{  Since  2.10}
\value{[\code{\link{GtkPageSetup}}]  a new \code{\link{GtkPageSetup}}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
