\alias{gtkPageSetupSetOrientation}
\name{gtkPageSetupSetOrientation}
\title{gtkPageSetupSetOrientation}
\description{Sets the page orientation of the \code{\link{GtkPageSetup}}.}
\usage{gtkPageSetupSetOrientation(object, orientation)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPageSetup}}]  a \code{\link{GtkPageSetup}}}
\item{\code{orientation}}{[\code{\link{GtkPageOrientation}}]  a \code{\link{GtkPageOrientation}} value}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
