\alias{gtkPageSetupSetRightMargin}
\name{gtkPageSetupSetRightMargin}
\title{gtkPageSetupSetRightMargin}
\description{Sets the right margin of the \verb{\link{GtkPageSetup}}.}
\usage{gtkPageSetupSetRightMargin(object, margin, unit)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPageSetup}}]  a \verb{\link{GtkPageSetup}}}
\item{\verb{margin}}{[numeric]  the new right margin in units of \code{unit}}
\item{\verb{unit}}{[\code{\link{GtkUnit}}]  the units for \code{margin}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
