\alias{gtkPageSetupSetTopMargin}
\name{gtkPageSetupSetTopMargin}
\title{gtkPageSetupSetTopMargin}
\description{Sets the top margin of the \verb{\link{GtkPageSetup}}.}
\usage{gtkPageSetupSetTopMargin(object, margin, unit)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPageSetup}}]  a \verb{\link{GtkPageSetup}}}
\item{\verb{margin}}{[numeric]  the new top margin in units of \code{unit}}
\item{\verb{unit}}{[\code{\link{GtkUnit}}]  the units for \code{margin}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
