\alias{gtkPaintLayout}
\name{gtkPaintLayout}
\title{gtkPaintLayout}
\description{\emph{undocumented
}}
\usage{gtkPaintLayout(object, window, state.type, use.text, area = NULL, widget = NULL, detail = NULL, x, y, layout)}
\arguments{
\item{\code{object}}{[\code{\link{GtkStyle}}] }
\item{\code{window}}{[\code{\link{GdkWindow}}] }
\item{\code{state.type}}{[\code{\link{GtkStateType}}] }
\item{\code{use.text}}{[logical] }
\item{\code{area}}{[\code{\link{GdkRectangle}}] }
\item{\code{widget}}{[\code{\link{GtkWidget}}] }
\item{\code{detail}}{[character] }
\item{\code{x}}{[integer] }
\item{\code{y}}{[integer] }
\item{\code{layout}}{[\code{\link{PangoLayout}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
