\alias{gtkPanedAdd1}
\name{gtkPanedAdd1}
\title{gtkPanedAdd1}
\description{Adds a child to the top or left pane with
default parameters. This is equivalent
to \code{gtk_paned_pack1 (paned, child, FALSE, TRUE)}.}
\usage{gtkPanedAdd1(object, child)}
\arguments{
\item{\verb{object}}{a paned widget}
\item{\verb{child}}{the child to add}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
