\alias{gtkPanedGetChild1}
\name{gtkPanedGetChild1}
\title{gtkPanedGetChild1}
\description{Obtains the first child of the paned widget.}
\usage{gtkPanedGetChild1(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPaned}} widget}}
\details{Since 2.4}
\value{[\code{\link{GtkWidget}}] first child, or \code{NULL} if it is not set.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
