\alias{gtkPanedGetChild2}
\name{gtkPanedGetChild2}
\title{gtkPanedGetChild2}
\description{Obtains the second child of the paned widget.}
\usage{gtkPanedGetChild2(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPaned}} widget}}
\details{Since 2.4}
\value{[\code{\link{GtkWidget}}] second child, or \code{NULL} if it is not set.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
