\alias{gtkPanedGetPosition}
\name{gtkPanedGetPosition}
\title{gtkPanedGetPosition}
\description{Obtains the position of the divider between the two panes.}
\usage{gtkPanedGetPosition(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPaned}} widget}}
\value{[integer] position of the divider}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
