\alias{gtkPanedPack1}
\name{gtkPanedPack1}
\title{gtkPanedPack1}
\description{Adds a child to the top or left pane.}
\usage{gtkPanedPack1(object, child, resize = FALSE, shrink = TRUE)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPaned}}] a paned widget}
\item{\code{child}}{[\code{\link{GtkWidget}}] the child to add}
\item{\code{resize}}{[logical] should this child expand when the paned widget is resized.}
\item{\code{shrink}}{[logical] can this child be made smaller than its requisition.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
