\alias{gtkPanedPack2}
\name{gtkPanedPack2}
\title{gtkPanedPack2}
\description{Adds a child to the bottom or right pane.}
\usage{gtkPanedPack2(object, child, resize = TRUE, shrink = TRUE)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPaned}}] a paned widget}
\item{\code{child}}{[\code{\link{GtkWidget}}] the child to add}
\item{\code{resize}}{[logical] should this child expand when the paned widget is resized.}
\item{\code{shrink}}{[logical] can this child be made smaller than its requisition.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
