\alias{gtkPanedSetPosition}
\name{gtkPanedSetPosition}
\title{gtkPanedSetPosition}
\description{Sets the position of the divider between the two panes.}
\usage{gtkPanedSetPosition(object, position)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPaned}}]  a \code{\link{GtkPaned}} widget}
\item{\code{position}}{[integer]  pixel position of divider, a negative value means that the position
           is unset.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
