\alias{gtkPaperSizeCopy}
\name{gtkPaperSizeCopy}
\title{gtkPaperSizeCopy}
\description{Copies an existing \verb{\link{GtkPaperSize}}.}
\usage{gtkPaperSizeCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkPaperSize}}]  a \verb{\link{GtkPaperSize}}}}
\details{  Since  2.10}
\value{[\verb{\link{GtkPaperSize}}]  a copy of \code{other}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
