\alias{gtkPaperSizeFree}
\name{gtkPaperSizeFree}
\title{gtkPaperSizeFree}
\description{Free the given \code{\link{GtkPaperSize}} object.}
\usage{gtkPaperSizeFree(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkPaperSize}}]  a \code{\link{GtkPaperSize}}}}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
