\alias{gtkPaperSizeGetDefault}
\name{gtkPaperSizeGetDefault}
\title{gtkPaperSizeGetDefault}
\description{Returns the name of the default paper size, which 
depends on the current locale.}
\usage{gtkPaperSizeGetDefault()}
\details{Since 2.10}
\value{[character] the name of the default paper size.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
