\alias{gtkPaperSizeGetDefaultBottomMargin}
\name{gtkPaperSizeGetDefaultBottomMargin}
\title{gtkPaperSizeGetDefaultBottomMargin}
\description{Gets the default bottom margin for the \verb{\link{GtkPaperSize}}.}
\usage{gtkPaperSizeGetDefaultBottomMargin(object, unit)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPaperSize}}]  a \verb{\link{GtkPaperSize}} object}
\item{\verb{unit}}{[\code{\link{GtkUnit}}]  the unit for the return value}
}
\details{  Since  2.10}
\value{[numeric]  the default bottom margin}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
