\alias{gtkPaperSizeGetDefaultLeftMargin}
\name{gtkPaperSizeGetDefaultLeftMargin}
\title{gtkPaperSizeGetDefaultLeftMargin}
\description{Gets the default left margin for the \verb{\link{GtkPaperSize}}.}
\usage{gtkPaperSizeGetDefaultLeftMargin(object, unit)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPaperSize}}]  a \verb{\link{GtkPaperSize}} object}
\item{\verb{unit}}{[\code{\link{GtkUnit}}]  the unit for the return value}
}
\details{  Since  2.10}
\value{[numeric]  the default left margin}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
