\alias{gtkPaperSizeGetDefaultRightMargin}
\name{gtkPaperSizeGetDefaultRightMargin}
\title{gtkPaperSizeGetDefaultRightMargin}
\description{Gets the default right margin for the \verb{\link{GtkPaperSize}}.}
\usage{gtkPaperSizeGetDefaultRightMargin(object, unit)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPaperSize}}]  a \verb{\link{GtkPaperSize}} object}
\item{\verb{unit}}{[\code{\link{GtkUnit}}]  the unit for the return value}
}
\details{  Since  2.10}
\value{[numeric]  the default right margin}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
