\alias{gtkPaperSizeGetDefaultTopMargin}
\name{gtkPaperSizeGetDefaultTopMargin}
\title{gtkPaperSizeGetDefaultTopMargin}
\description{Gets the default top margin for the \verb{\link{GtkPaperSize}}.}
\usage{gtkPaperSizeGetDefaultTopMargin(object, unit)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPaperSize}}]  a \verb{\link{GtkPaperSize}} object}
\item{\verb{unit}}{[\code{\link{GtkUnit}}]  the unit for the return value}
}
\details{  Since  2.10}
\value{[numeric]  the default top margin}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
