\alias{gtkPaperSizeGetDisplayName}
\name{gtkPaperSizeGetDisplayName}
\title{gtkPaperSizeGetDisplayName}
\description{Gets the human-readable name of the \code{\link{GtkPaperSize}}.}
\usage{gtkPaperSizeGetDisplayName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPaperSize}} object}}
\details{Since 2.10}
\value{[character] the human-readable name of \code{size}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
