\alias{gtkPaperSizeGetHeight}
\name{gtkPaperSizeGetHeight}
\title{gtkPaperSizeGetHeight}
\description{Gets the paper height of the \code{\link{GtkPaperSize}}, in 
units of \code{unit}.}
\usage{gtkPaperSizeGetHeight(object, unit)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPaperSize}} object}
\item{\verb{unit}}{the unit for the return value}
}
\details{Since 2.10}
\value{[numeric] the paper height}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
