\alias{gtkPaperSizeGetName}
\name{gtkPaperSizeGetName}
\title{gtkPaperSizeGetName}
\description{Gets the name of the \verb{\link{GtkPaperSize}}.}
\usage{gtkPaperSizeGetName(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkPaperSize}}]  a \verb{\link{GtkPaperSize}} object}}
\details{  Since  2.10}
\value{[character]  the name of \code{size}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
