\alias{gtkPaperSizeGetPaperSizes}
\name{gtkPaperSizeGetPaperSizes}
\title{gtkPaperSizeGetPaperSizes}
\description{Creates a list of known paper sizes.}
\usage{gtkPaperSizeGetPaperSizes(include.custom)}
\arguments{\item{\code{include.custom}}{[logical]  whether to include custom paper sizes
    as defined in the page setup dialog}}
\details{  Since  2.12}
\value{[list]  a newly allocated list of newly 
   allocated \code{\link{GtkPaperSize}} objects}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
