\alias{gtkPaperSizeGetPpdName}
\name{gtkPaperSizeGetPpdName}
\title{gtkPaperSizeGetPpdName}
\description{Gets the PPD name of the \verb{\link{GtkPaperSize}}, which
may be \code{NULL}.}
\usage{gtkPaperSizeGetPpdName(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkPaperSize}}]  a \verb{\link{GtkPaperSize}} object}}
\details{  Since  2.10}
\value{[character]  the PPD name of \code{size}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
