\alias{gtkPaperSizeGetWidth}
\name{gtkPaperSizeGetWidth}
\title{gtkPaperSizeGetWidth}
\description{Gets the paper width of the \code{\link{GtkPaperSize}}, in 
units of \code{unit}.}
\usage{gtkPaperSizeGetWidth(object, unit)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPaperSize}}]  a \code{\link{GtkPaperSize}} object}
\item{\code{unit}}{[\code{\link{GtkUnit}}]  the unit for the return value}
}
\details{  Since  2.10}
\value{[numeric]  the paper width }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
