\alias{gtkPaperSizeIsCustom}
\name{gtkPaperSizeIsCustom}
\title{gtkPaperSizeIsCustom}
\description{Returns \code{TRUE} if \code{size} is not a standard paper size.}
\usage{gtkPaperSizeIsCustom(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkPaperSize}}]  a \code{\link{GtkPaperSize}} object}}
\value{[logical]  whether \code{size} is a custom paper size.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
