\alias{gtkPaperSizeIsEqual}
\name{gtkPaperSizeIsEqual}
\title{gtkPaperSizeIsEqual}
\description{Compares two \verb{\link{GtkPaperSize}} objects.}
\usage{gtkPaperSizeIsEqual(object, size2)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPaperSize}}]  a \verb{\link{GtkPaperSize}} object}
\item{\verb{size2}}{[\verb{\link{GtkPaperSize}}]  another \verb{\link{GtkPaperSize}} object}
}
\details{  Since  2.10}
\value{[logical] \code{TRUE}, if \code{size1} and \code{size2} 
represent the same paper size}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
