\alias{gtkPaperSizeSetSize}
\name{gtkPaperSizeSetSize}
\title{gtkPaperSizeSetSize}
\description{Changes the dimensions of a \code{size} to \code{width} x \code{height}.}
\usage{gtkPaperSizeSetSize(object, width, height, unit)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPaperSize}}]  a custom \verb{\link{GtkPaperSize}} object}
\item{\verb{width}}{[numeric]  the new width in units of \code{unit}}
\item{\verb{height}}{[numeric]  the new height in units of \code{unit}}
\item{\verb{unit}}{[\code{\link{GtkUnit}}]  the unit for \code{width} and \code{height}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
