\alias{gtkPlugGetEmbedded}
\name{gtkPlugGetEmbedded}
\title{gtkPlugGetEmbedded}
\description{Determines whether the plug is embedded in a socket.}
\usage{gtkPlugGetEmbedded(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPlug}}}}
\details{Since 2.14}
\value{[logical] \code{TRUE} if the plug is embedded in a socket}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
