\alias{gtkPlugGetSocketWindow}
\name{gtkPlugGetSocketWindow}
\title{gtkPlugGetSocketWindow}
\description{Retrieves the socket the plug is embedded in.}
\usage{gtkPlugGetSocketWindow(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPlug}}}}
\details{Since 2.14}
\value{[\code{\link{GdkWindow}}] the window of the socket, or \code{NULL}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
