\alias{gtkPreviewGetCmap}
\name{gtkPreviewGetCmap}
\title{gtkPreviewGetCmap}
\description{
Returns the colormap used by preview widgets. This
function is deprecated, and you should use
\code{\link{gdkRgbGetCmap}} instead.
\strong{WARNING: \code{gtk_preview_get_cmap} is deprecated and should not be used in newly-written code.}
}
\usage{gtkPreviewGetCmap()}
\value{[\code{\link{GdkColormap}}] the colormap for previews.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
