\alias{gtkPreviewGetInfo}
\name{gtkPreviewGetInfo}
\title{gtkPreviewGetInfo}
\description{
Return a \verb{\link{GtkPreviewInfo}} structure containing 
global information about preview widgets.
\strong{WARNING: \code{gtk_preview_get_info} is deprecated and should not be used in newly-written code.}
}
\usage{gtkPreviewGetInfo()}
\value{[\verb{\link{GtkPreviewInfo}}] a \verb{\link{GtkPreviewInfo}} structure.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
