\alias{gtkPreviewGetVisual}
\name{gtkPreviewGetVisual}
\title{gtkPreviewGetVisual}
\description{
Returns the visual used by preview widgets. This
function is deprecated, and you should use
\code{\link{gdkRgbGetVisual}} instead.
\strong{WARNING: \code{gtk_preview_get_visual} is deprecated and should not be used in newly-written code.}
}
\usage{gtkPreviewGetVisual()}
\value{[\code{\link{GdkVisual}}] the visual for previews.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
