\alias{gtkPreviewNew}
\name{gtkPreviewNew}
\title{gtkPreviewNew}
\description{
Create a new preview widget.
\strong{WARNING: \code{gtk_preview_new} is deprecated and should not be used in newly-written code.  
Create a new preview widget.}
}
\usage{gtkPreviewNew(type, show = TRUE)}
\arguments{\item{\code{type}}{[\code{\link{GtkPreviewType}}] the type data contained by the widget. 
(Grayscale or RGB)}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
