\alias{gtkPreviewReset}
\name{gtkPreviewReset}
\title{gtkPreviewReset}
\description{
This function is deprecated and does nothing. It was
once used for changing the colormap and visual on the fly.
\strong{WARNING: \code{gtk_preview_reset} is deprecated and should not be used in newly-written code.}
}
\usage{gtkPreviewReset()}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
