\alias{gtkPreviewSetColorCube}
\name{gtkPreviewSetColorCube}
\title{gtkPreviewSetColorCube}
\description{
This function is deprecated and does nothing. GdkRGB
automatically picks an optimium color cube for the
display.
\strong{WARNING: \code{gtk_preview_set_color_cube} is deprecated and should not be used in newly-written code.}
}
\usage{gtkPreviewSetColorCube(nred.shades, ngreen.shades, nblue.shades, ngray.shades)}
\arguments{
\item{\verb{nred.shades}}{[numeric] ignored}
\item{\verb{ngreen.shades}}{[numeric] ignored}
\item{\verb{nblue.shades}}{[numeric] ignored}
\item{\verb{ngray.shades}}{[numeric] ignored}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
