\alias{gtkPreviewSetDither}
\name{gtkPreviewSetDither}
\title{gtkPreviewSetDither}
\description{
Set the dithering mode for the display. 
\strong{WARNING: \code{gtk_preview_set_dither} is deprecated and should not be used in newly-written code.}
}
\usage{gtkPreviewSetDither(object, dither)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPreview}}] a \verb{\link{GtkPreview}}.}
\item{\verb{dither}}{[\code{\link{GdkRgbDither}}] the dithering mode.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
