\alias{gtkPreviewUninit}
\name{gtkPreviewUninit}
\title{gtkPreviewUninit}
\description{
This function is deprecated and does nothing.
\strong{WARNING: \code{gtk_preview_uninit} is deprecated and should not be used in newly-written code.}
}
\usage{gtkPreviewUninit()}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
