\alias{gtkPrintContextCreatePangoContext}
\name{gtkPrintContextCreatePangoContext}
\title{gtkPrintContextCreatePangoContext}
\description{Creates a new \verb{\link{PangoContext}} that can be used with the
\verb{\link{GtkPrintContext}}.}
\usage{gtkPrintContextCreatePangoContext(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkPrintContext}}]  a \verb{\link{GtkPrintContext}} }}
\details{  Since  2.10}
\value{[\verb{\link{PangoContext}}]  a new Pango context for \code{context}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
