\alias{gtkPrintContextCreatePangoLayout}
\name{gtkPrintContextCreatePangoLayout}
\title{gtkPrintContextCreatePangoLayout}
\description{Creates a new \verb{\link{PangoLayout}} that is suitable for use
with the \verb{\link{GtkPrintContext}}.}
\usage{gtkPrintContextCreatePangoLayout(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkPrintContext}}]  a \verb{\link{GtkPrintContext}}}}
\details{  Since  2.10}
\value{[\verb{\link{PangoLayout}}]  a new Pango layout for \code{context}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
