\alias{gtkPrintContextGetCairoContext}
\name{gtkPrintContextGetCairoContext}
\title{gtkPrintContextGetCairoContext}
\description{Obtains the cairo context that is associated with the
\code{\link{GtkPrintContext}}.}
\usage{gtkPrintContextGetCairoContext(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkPrintContext}}]  a \code{\link{GtkPrintContext}}}}
\details{  Since  2.10}
\value{[\code{\link{Cairo}}]  the cairo context of \code{context}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
