\alias{gtkPrintContextGetDpiX}
\name{gtkPrintContextGetDpiX}
\title{gtkPrintContextGetDpiX}
\description{Obtains the horizontal resolution of the \verb{\link{GtkPrintContext}},
in dots per inch.}
\usage{gtkPrintContextGetDpiX(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkPrintContext}}]  a \verb{\link{GtkPrintContext}}}}
\details{  Since  2.10}
\value{[numeric]  the horizontal resolution of \code{context}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
