\alias{gtkPrintContextGetPageSetup}
\name{gtkPrintContextGetPageSetup}
\title{gtkPrintContextGetPageSetup}
\description{Obtains the \verb{\link{GtkPageSetup}} that determines the page
dimensions of the \verb{\link{GtkPrintContext}}.}
\usage{gtkPrintContextGetPageSetup(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkPrintContext}}]  a \verb{\link{GtkPrintContext}}}}
\details{  Since  2.10}
\value{[\verb{\link{GtkPageSetup}}]  the page setup of \code{context}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
