\alias{gtkPrintContextGetPangoFontmap}
\name{gtkPrintContextGetPangoFontmap}
\title{gtkPrintContextGetPangoFontmap}
\description{Returns a \verb{\link{PangoFontMap}} that is suitable for use 
with the \verb{\link{GtkPrintContext}}.}
\usage{gtkPrintContextGetPangoFontmap(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkPrintContext}}]  a \verb{\link{GtkPrintContext}}}}
\details{  Since  2.10}
\value{[\verb{\link{PangoFontMap}}]  the font map of \code{context}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
