\alias{gtkPrintContextGetWidth}
\name{gtkPrintContextGetWidth}
\title{gtkPrintContextGetWidth}
\description{Obtains the width of the \verb{\link{GtkPrintContext}}, in pixels.}
\usage{gtkPrintContextGetWidth(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkPrintContext}}]  a \verb{\link{GtkPrintContext}}}}
\details{  Since  2.10 }
\value{[numeric]  the width of \code{context}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
