\alias{gtkPrintOperationCancel}
\name{gtkPrintOperationCancel}
\title{gtkPrintOperationCancel}
\description{Cancels a running print operation. This function may
be called from a \verb{"begin-print"}, 
\verb{"paginate"} or \verb{"draw-page"}
signal handler to stop the currently running print 
operation.}
\usage{gtkPrintOperationCancel(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintOperation}}}}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
