\alias{gtkPrintOperationGetEmbedPageSetup}
\name{gtkPrintOperationGetEmbedPageSetup}
\title{gtkPrintOperationGetEmbedPageSetup}
\description{Gets the value of \verb{"embed-page-setup"} property.}
\usage{gtkPrintOperationGetEmbedPageSetup(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintOperation}}}}
\details{Since 2.18}
\value{[logical] whether page setup selection combos are embedded}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
