\alias{gtkPrintOperationGetHasSelection}
\name{gtkPrintOperationGetHasSelection}
\title{gtkPrintOperationGetHasSelection}
\description{Gets the value of \verb{"has-selection"} property.}
\usage{gtkPrintOperationGetHasSelection(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintOperation}}}}
\details{Since 2.18}
\value{[logical] whether there is a selection}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
