\alias{gtkPrintOperationGetStatus}
\name{gtkPrintOperationGetStatus}
\title{gtkPrintOperationGetStatus}
\description{Returns the status of the print operation. 
Also see \code{\link{gtkPrintOperationGetStatusString}}.}
\usage{gtkPrintOperationGetStatus(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintOperation}}}}
\details{Since 2.10}
\value{[\code{\link{GtkPrintStatus}}] the status of the print operation}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
