\alias{gtkPrintOperationGetSupportSelection}
\name{gtkPrintOperationGetSupportSelection}
\title{gtkPrintOperationGetSupportSelection}
\description{Gets the value of \verb{"support-selection"} property.}
\usage{gtkPrintOperationGetSupportSelection(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintOperation}}}}
\details{Since 2.18}
\value{[logical] whether the application supports print of selection}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
