\alias{gtkPrintOperationNew}
\name{gtkPrintOperationNew}
\title{gtkPrintOperationNew}
\description{Creates a new \verb{\link{GtkPrintOperation}}.}
\usage{gtkPrintOperationNew()}
\details{  Since  2.10}
\value{[\verb{\link{GtkPrintOperation}}]  a new \verb{\link{GtkPrintOperation}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
