\alias{gtkPrintOperationPreviewEndPreview}
\name{gtkPrintOperationPreviewEndPreview}
\title{gtkPrintOperationPreviewEndPreview}
\description{Ends a preview. }
\usage{gtkPrintOperationPreviewEndPreview(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintOperationPreview}}}}
\details{This function must be called to finish a custom print preview.
  Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
