\alias{gtkPrintOperationPreviewIsSelected}
\name{gtkPrintOperationPreviewIsSelected}
\title{gtkPrintOperationPreviewIsSelected}
\description{Returns whether the given page is included in the set of pages that
have been selected for printing.}
\usage{gtkPrintOperationPreviewIsSelected(object, page.nr)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintOperationPreview}}}
\item{\verb{page.nr}}{a page number}
}
\details{Since 2.10}
\value{[logical] \code{TRUE} if the page has been selected for printing}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
