\alias{gtkPrintOperationSetCustomTabLabel}
\name{gtkPrintOperationSetCustomTabLabel}
\title{gtkPrintOperationSetCustomTabLabel}
\description{Sets the label for the tab holding custom widgets.}
\usage{gtkPrintOperationSetCustomTabLabel(object, label)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintOperation}}]  a \code{\link{GtkPrintOperation}}}
\item{\code{label}}{[character]  the label to use, or \code{NULL} to use the default label}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
