\alias{gtkPrintOperationSetEmbedPageSetup}
\name{gtkPrintOperationSetEmbedPageSetup}
\title{gtkPrintOperationSetEmbedPageSetup}
\description{Embed page size combo box and orientation combo box into page setup page.
Selected page setup is stored as default page setup in \code{\link{GtkPrintOperation}}.}
\usage{gtkPrintOperationSetEmbedPageSetup(object, embed)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintOperation}}}
\item{\verb{embed}}{\code{TRUE} to embed page setup selection in the \verb{GtkPrintDialog}}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
