\alias{gtkPrintOperationSetSupportSelection}
\name{gtkPrintOperationSetSupportSelection}
\title{gtkPrintOperationSetSupportSelection}
\description{Sets whether selection is supported by \code{\link{GtkPrintOperation}}.}
\usage{gtkPrintOperationSetSupportSelection(object, support.selection)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintOperation}}}
\item{\verb{support.selection}}{\code{TRUE} to support selection}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
