\alias{gtkPrintSettingsCopy}
\name{gtkPrintSettingsCopy}
\title{gtkPrintSettingsCopy}
\description{Copies a \verb{\link{GtkPrintSettings}} object.}
\usage{gtkPrintSettingsCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkPrintSettings}}]  a \verb{\link{GtkPrintSettings}}}}
\details{  Since  2.10}
\value{[\verb{\link{GtkPrintSettings}}]  a newly allocated copy of \code{other}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
