\alias{gtkPrintSettingsForeach}
\name{gtkPrintSettingsForeach}
\title{gtkPrintSettingsForeach}
\description{Calls \code{func} for each key-value pair of \code{settings}.}
\usage{gtkPrintSettingsForeach(object, func, user.data = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPrintSettings}}]  a \verb{\link{GtkPrintSettings}}}
\item{\verb{func}}{[\verb{\link{GtkPrintSettingsFunc}}]  the function to call}
\item{\verb{user.data}}{[R object]  user data for \code{func}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
