\alias{gtkPrintSettingsGet}
\name{gtkPrintSettingsGet}
\title{gtkPrintSettingsGet}
\description{Looks up the string value associated with \code{key}.}
\usage{gtkPrintSettingsGet(object, key)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{key}}{[character]  a key}
}
\details{  Since  2.10}
\value{[character]  the string value for \code{key}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
