\alias{gtkPrintSettingsGetCollate}
\name{gtkPrintSettingsGetCollate}
\title{gtkPrintSettingsGetCollate}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_COLLATE}.}
\usage{gtkPrintSettingsGetCollate(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.10}
\value{[logical] whether to collate the printed pages}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
