\alias{gtkPrintSettingsGetDefaultSource}
\name{gtkPrintSettingsGetDefaultSource}
\title{gtkPrintSettingsGetDefaultSource}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_DEFAULT_SOURCE}.}
\usage{gtkPrintSettingsGetDefaultSource(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}}
\details{  Since  2.10}
\value{[character]  the default source}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
