\alias{gtkPrintSettingsGetDither}
\name{gtkPrintSettingsGetDither}
\title{gtkPrintSettingsGetDither}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_DITHER}.}
\usage{gtkPrintSettingsGetDither(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.10}
\value{[character] the dithering that is used}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
