\alias{gtkPrintSettingsGetDuplex}
\name{gtkPrintSettingsGetDuplex}
\title{gtkPrintSettingsGetDuplex}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_DUPLEX}.}
\usage{gtkPrintSettingsGetDuplex(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.10}
\value{[\code{\link{GtkPrintDuplex}}] whether to print the output in duplex.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
