\alias{gtkPrintSettingsGetFinishings}
\name{gtkPrintSettingsGetFinishings}
\title{gtkPrintSettingsGetFinishings}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_FINISHINGS}.}
\usage{gtkPrintSettingsGetFinishings(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}}
\details{  Since  2.10}
\value{[character]  the finishings}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
