\alias{gtkPrintSettingsGetInt}
\name{gtkPrintSettingsGetInt}
\title{gtkPrintSettingsGetInt}
\description{Returns the integer value of \code{key}, or 0.}
\usage{gtkPrintSettingsGetInt(object, key)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{key}}{[character]  a key}
}
\details{  Since  2.10}
\value{[integer]  the integer value of \code{key} }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
