\alias{gtkPrintSettingsGetMediaType}
\name{gtkPrintSettingsGetMediaType}
\title{gtkPrintSettingsGetMediaType}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_MEDIA_TYPE}.}
\usage{gtkPrintSettingsGetMediaType(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{The set of media types is defined in PWG 5101.1-2002 PWG.

  Since 2.10}
\value{[character] the media type}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
