\alias{gtkPrintSettingsGetNCopies}
\name{gtkPrintSettingsGetNCopies}
\title{gtkPrintSettingsGetNCopies}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_N_COPIES}.}
\usage{gtkPrintSettingsGetNCopies(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.10}
\value{[integer] the number of copies to print}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
