\alias{gtkPrintSettingsGetNumberUp}
\name{gtkPrintSettingsGetNumberUp}
\title{gtkPrintSettingsGetNumberUp}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_NUMBER_UP}.}
\usage{gtkPrintSettingsGetNumberUp(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.10}
\value{[integer] the number of pages per sheet}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
