\alias{gtkPrintSettingsGetNumberUpLayout}
\name{gtkPrintSettingsGetNumberUpLayout}
\title{gtkPrintSettingsGetNumberUpLayout}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_NUMBER_UP_LAYOUT}.}
\usage{gtkPrintSettingsGetNumberUpLayout(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.14}
\value{[\code{\link{GtkNumberUpLayout}}] layout of page in number-up mode}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
