\alias{gtkPrintSettingsGetOutputBin}
\name{gtkPrintSettingsGetOutputBin}
\title{gtkPrintSettingsGetOutputBin}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_OUTPUT_BIN}.}
\usage{gtkPrintSettingsGetOutputBin(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.10}
\value{[character] the output bin}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
