\alias{gtkPrintSettingsGetPageRanges}
\name{gtkPrintSettingsGetPageRanges}
\title{gtkPrintSettingsGetPageRanges}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_PAGE_RANGES}.}
\usage{gtkPrintSettingsGetPageRanges(object, num.ranges)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{num.ranges}}{[integer]  return location for the length of the returned list}
}
\details{  Since  2.10}
\value{[\code{\link{GtkPageRange}}]  a list of \code{\link{GtkPageRange}}s. }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
