\alias{gtkPrintSettingsGetPageSet}
\name{gtkPrintSettingsGetPageSet}
\title{gtkPrintSettingsGetPageSet}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_PAGE_SET}.}
\usage{gtkPrintSettingsGetPageSet(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.10}
\value{[\code{\link{GtkPageSet}}] the set of pages to print}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
