\alias{gtkPrintSettingsGetPaperWidth}
\name{gtkPrintSettingsGetPaperWidth}
\title{gtkPrintSettingsGetPaperWidth}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_PAPER_WIDTH},
converted to \code{unit}.}
\usage{gtkPrintSettingsGetPaperWidth(object, unit)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{unit}}{the unit for the return value}
}
\details{Since 2.10}
\value{[numeric] the paper width, in units of \code{unit}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
