\alias{gtkPrintSettingsGetPrinter}
\name{gtkPrintSettingsGetPrinter}
\title{gtkPrintSettingsGetPrinter}
\description{Convenience function to obtain the value of 
\code{GTK_PRINT_SETTINGS_PRINTER}.}
\usage{gtkPrintSettingsGetPrinter(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.10}
\value{[character] the printer name}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
