\alias{gtkPrintSettingsGetPrinterLpi}
\name{gtkPrintSettingsGetPrinterLpi}
\title{gtkPrintSettingsGetPrinterLpi}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_PRINTER_LPI}.}
\usage{gtkPrintSettingsGetPrinterLpi(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.16}
\value{[numeric] the resolution in lpi (lines per inch)}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
