\alias{gtkPrintSettingsGetPrintPages}
\name{gtkPrintSettingsGetPrintPages}
\title{gtkPrintSettingsGetPrintPages}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_PRINT_PAGES}.}
\usage{gtkPrintSettingsGetPrintPages(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.10}
\value{[\code{\link{GtkPrintPages}}] which pages to print}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
