\alias{gtkPrintSettingsGetResolution}
\name{gtkPrintSettingsGetResolution}
\title{gtkPrintSettingsGetResolution}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_RESOLUTION}.}
\usage{gtkPrintSettingsGetResolution(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.10}
\value{[integer] the resolution in dpi}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
