\alias{gtkPrintSettingsGetResolutionX}
\name{gtkPrintSettingsGetResolutionX}
\title{gtkPrintSettingsGetResolutionX}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_RESOLUTION_X}.}
\usage{gtkPrintSettingsGetResolutionX(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.16}
\value{[integer] the horizontal resolution in dpi}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
