\alias{gtkPrintSettingsGetResolutionY}
\name{gtkPrintSettingsGetResolutionY}
\title{gtkPrintSettingsGetResolutionY}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_RESOLUTION_Y}.}
\usage{gtkPrintSettingsGetResolutionY(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.16}
\value{[integer] the vertical resolution in dpi}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
