\alias{gtkPrintSettingsGetReverse}
\name{gtkPrintSettingsGetReverse}
\title{gtkPrintSettingsGetReverse}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_REVERSE}.}
\usage{gtkPrintSettingsGetReverse(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.10}
\value{[logical] whether to reverse the order of the printed pages}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
