\alias{gtkPrintSettingsGetScale}
\name{gtkPrintSettingsGetScale}
\title{gtkPrintSettingsGetScale}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_SCALE}.}
\usage{gtkPrintSettingsGetScale(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.10}
\value{[numeric] the scale in percent}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
