\alias{gtkPrintSettingsNew}
\name{gtkPrintSettingsNew}
\title{gtkPrintSettingsNew}
\description{Creates a new \verb{\link{GtkPrintSettings}} object.}
\usage{gtkPrintSettingsNew()}
\details{  Since  2.10}
\value{[\verb{\link{GtkPrintSettings}}]  a new \verb{\link{GtkPrintSettings}} object}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
