\alias{gtkPrintSettingsSet}
\name{gtkPrintSettingsSet}
\title{gtkPrintSettingsSet}
\description{Associates \code{value} with \code{key}.}
\usage{gtkPrintSettingsSet(object, key, value)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{key}}{[character]  a key}
\item{\code{value}}{[character]  a string value, or \code{NULL}}
}
\details{  Since  2.10 }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
