\alias{gtkPrintSettingsSetDither}
\name{gtkPrintSettingsSetDither}
\title{gtkPrintSettingsSetDither}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_DITHER}.}
\usage{gtkPrintSettingsSetDither(object, dither)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{dither}}{the dithering that is used}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
