\alias{gtkPrintSettingsSetDouble}
\name{gtkPrintSettingsSetDouble}
\title{gtkPrintSettingsSetDouble}
\description{Sets \code{key} to a double value.}
\usage{gtkPrintSettingsSetDouble(object, key, value)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPrintSettings}}]  a \verb{\link{GtkPrintSettings}}}
\item{\verb{key}}{[character]  a key }
\item{\verb{value}}{[numeric]  a double value}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
