\alias{gtkPrintSettingsSetFinishings}
\name{gtkPrintSettingsSetFinishings}
\title{gtkPrintSettingsSetFinishings}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_FINISHINGS}.}
\usage{gtkPrintSettingsSetFinishings(object, finishings)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{finishings}}{[character]  the finishings}
}
\details{  Since  2.10}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
