\alias{gtkPrintSettingsSetInt}
\name{gtkPrintSettingsSetInt}
\title{gtkPrintSettingsSetInt}
\description{Sets \code{key} to an integer value.}
\usage{gtkPrintSettingsSetInt(object, key, value)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{key}}{[character]  a key}
\item{\code{value}}{[integer]  an integer }
}
\details{  Since  2.10 }
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
