\alias{gtkPrintSettingsSetLength}
\name{gtkPrintSettingsSetLength}
\title{gtkPrintSettingsSetLength}
\description{Associates a length in units of \code{unit} with \code{key}.}
\usage{gtkPrintSettingsSetLength(object, key, value, unit)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPrintSettings}}]  a \verb{\link{GtkPrintSettings}}}
\item{\verb{key}}{[character]  a key}
\item{\verb{value}}{[numeric]  a length}
\item{\verb{unit}}{[\code{\link{GtkUnit}}]  the unit of \code{length}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
