\alias{gtkPrintSettingsSetMediaType}
\name{gtkPrintSettingsSetMediaType}
\title{gtkPrintSettingsSetMediaType}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_MEDIA_TYPE}.}
\usage{gtkPrintSettingsSetMediaType(object, media.type)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{media.type}}{[character]  the media type}
}
\details{The set of media types is defined in PWG 5101.1-2002 PWG.
  

  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
