\alias{gtkPrintSettingsSetNCopies}
\name{gtkPrintSettingsSetNCopies}
\title{gtkPrintSettingsSetNCopies}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_N_COPIES}.}
\usage{gtkPrintSettingsSetNCopies(object, num.copies)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{num.copies}}{[integer]  the number of copies }
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
