\alias{gtkPrintSettingsSetNumberUp}
\name{gtkPrintSettingsSetNumberUp}
\title{gtkPrintSettingsSetNumberUp}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_NUMBER_UP}.}
\usage{gtkPrintSettingsSetNumberUp(object, number.up)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{number.up}}{[integer]  the number of pages per sheet }
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
