\alias{gtkPrintSettingsSetNumberUpLayout}
\name{gtkPrintSettingsSetNumberUpLayout}
\title{gtkPrintSettingsSetNumberUpLayout}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_NUMBER_UP_LAYOUT}.}
\usage{gtkPrintSettingsSetNumberUpLayout(object, number.up.layout)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{number.up.layout}}{a \code{\link{GtkNumberUpLayout}} value}
}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
