\alias{gtkPrintSettingsSetOutputBin}
\name{gtkPrintSettingsSetOutputBin}
\title{gtkPrintSettingsSetOutputBin}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_OUTPUT_BIN}.}
\usage{gtkPrintSettingsSetOutputBin(object, output.bin)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{output.bin}}{[character]  the output bin}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
