\alias{gtkPrintSettingsSetPageRanges}
\name{gtkPrintSettingsSetPageRanges}
\title{gtkPrintSettingsSetPageRanges}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_PAGE_RANGES}.}
\usage{gtkPrintSettingsSetPageRanges(object, page.ranges, num.ranges)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{page.ranges}}{[\code{\link{GtkPageRange}}]  a list of \code{\link{GtkPageRange}}s}
\item{\code{num.ranges}}{[integer]  the length of \code{page.ranges}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
