\alias{gtkPrintSettingsSetPaperHeight}
\name{gtkPrintSettingsSetPaperHeight}
\title{gtkPrintSettingsSetPaperHeight}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_PAPER_HEIGHT}.}
\usage{gtkPrintSettingsSetPaperHeight(object, height, unit)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPrintSettings}}]  a \verb{\link{GtkPrintSettings}}}
\item{\verb{height}}{[numeric]  the paper height}
\item{\verb{unit}}{[\code{\link{GtkUnit}}]  the units of \code{height}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
