\alias{gtkPrintSettingsSetPaperSize}
\name{gtkPrintSettingsSetPaperSize}
\title{gtkPrintSettingsSetPaperSize}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_PAPER_FORMAT},
\code{GTK_PRINT_SETTINGS_PAPER_WIDTH} and
\code{GTK_PRINT_SETTINGS_PAPER_HEIGHT}.}
\usage{gtkPrintSettingsSetPaperSize(object, paper.size)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{paper.size}}{[\code{\link{GtkPaperSize}}]  a paper size}
}
\details{  Since  2.10}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
