\alias{gtkPrintSettingsSetPrinterLpi}
\name{gtkPrintSettingsSetPrinterLpi}
\title{gtkPrintSettingsSetPrinterLpi}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_PRINTER_LPI}.}
\usage{gtkPrintSettingsSetPrinterLpi(object, lpi)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{lpi}}{the resolution in lpi (lines per inch)}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
