\alias{gtkPrintSettingsSetQuality}
\name{gtkPrintSettingsSetQuality}
\title{gtkPrintSettingsSetQuality}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_QUALITY}.}
\usage{gtkPrintSettingsSetQuality(object, quality)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{quality}}{[\code{\link{GtkPrintQuality}}]  a \code{\link{GtkPrintQuality}} value}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
