\alias{gtkPrintSettingsSetResolution}
\name{gtkPrintSettingsSetResolution}
\title{gtkPrintSettingsSetResolution}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_RESOLUTION}.}
\usage{gtkPrintSettingsSetResolution(object, resolution)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{resolution}}{[integer]  the resolution in dpi}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
