\alias{gtkPrintSettingsSetResolutionXy}
\name{gtkPrintSettingsSetResolutionXy}
\title{gtkPrintSettingsSetResolutionXy}
\description{Sets the values of \code{GTK_PRINT_SETTINGS_RESOLUTION},
\code{GTK_PRINT_SETTINGS_RESOLUTION_X} and
\code{GTK_PRINT_SETTINGS_RESOLUTION_Y}.}
\usage{gtkPrintSettingsSetResolutionXy(object, resolution.x, resolution.y)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{resolution.x}}{the horizontal resolution in dpi}
\item{\verb{resolution.y}}{the vertical resolution in dpi}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
