\alias{gtkPrintSettingsSetScale}
\name{gtkPrintSettingsSetScale}
\title{gtkPrintSettingsSetScale}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_SCALE}.}
\usage{gtkPrintSettingsSetScale(object, scale)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{scale}}{[numeric]  the scale in percent}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
