\alias{gtkPrintSettingsSetUseColor}
\name{gtkPrintSettingsSetUseColor}
\title{gtkPrintSettingsSetUseColor}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_USE_COLOR}.}
\usage{gtkPrintSettingsSetUseColor(object, use.color)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{use.color}}{[logical]  whether to use color}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
