\alias{gtkPrintSettingsUnset}
\name{gtkPrintSettingsUnset}
\title{gtkPrintSettingsUnset}
\description{Removes any value associated with \code{key}. 
This has the same effect as setting the value to \code{NULL}.}
\usage{gtkPrintSettingsUnset(object, key)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{key}}{[character]  a key}
}
\details{  Since  2.10 }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
