\alias{gtkProgressBarGetEllipsize}
\name{gtkProgressBarGetEllipsize}
\title{gtkProgressBarGetEllipsize}
\description{Returns the ellipsizing position of the progressbar. 
See \code{\link{gtkProgressBarSetEllipsize}}.}
\usage{gtkProgressBarGetEllipsize(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkProgressBar}}]  a \code{\link{GtkProgressBar}}}}
\details{  Since  2.6}
\value{[\code{\link{PangoEllipsizeMode}}] \code{\link{PangoEllipsizeMode}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
