\alias{gtkProgressBarGetFraction}
\name{gtkProgressBarGetFraction}
\title{gtkProgressBarGetFraction}
\description{Returns the current fraction of the task that's been completed.}
\usage{gtkProgressBarGetFraction(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkProgressBar}}}}
\value{[numeric] a fraction from 0.0 to 1.0}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
