\alias{gtkProgressBarGetOrientation}
\name{gtkProgressBarGetOrientation}
\title{gtkProgressBarGetOrientation}
\description{Retrieves the current progress bar orientation.}
\usage{gtkProgressBarGetOrientation(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkProgressBar}}]  a \code{\link{GtkProgressBar}}}}
\value{[\code{\link{GtkProgressBarOrientation}}]  orientation of the progress bar}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
