\alias{gtkProgressBarGetPulseStep}
\name{gtkProgressBarGetPulseStep}
\title{gtkProgressBarGetPulseStep}
\description{Retrieves the pulse step set with \code{\link{gtkProgressBarSetPulseStep}}}
\usage{gtkProgressBarGetPulseStep(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkProgressBar}}}}
\value{[numeric] a fraction from 0.0 to 1.0}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
