\alias{gtkProgressBarGetText}
\name{gtkProgressBarGetText}
\title{gtkProgressBarGetText}
\description{Retrieves the text displayed superimposed on the progress bar,
if any, otherwise \code{NULL}. The return value is a reference
to the text, not a copy of it, so will become invalid
if you change the text in the progress bar.}
\usage{gtkProgressBarGetText(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkProgressBar}}}}
\value{[character] text,}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
