\alias{gtkProgressBarNewWithAdjustment}
\name{gtkProgressBarNewWithAdjustment}
\title{gtkProgressBarNewWithAdjustment}
\description{
Creates a new \verb{\link{GtkProgressBar}} with an associated \verb{\link{GtkAdjustment}}.
\strong{WARNING: \code{gtk_progress_bar_new_with_adjustment} is deprecated and should not be used in newly-written code.}
}
\usage{gtkProgressBarNewWithAdjustment(adjustment = NULL, show = TRUE)}
\arguments{\item{\verb{adjustment}}{[\verb{\link{GtkAdjustment}}] a \verb{\link{GtkAdjustment}}.}}
\value{[\verb{\link{GtkWidget}}] a \verb{\link{GtkProgressBar}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
