\alias{gtkProgressBarSetDiscreteBlocks}
\name{gtkProgressBarSetDiscreteBlocks}
\title{gtkProgressBarSetDiscreteBlocks}
\description{
Sets the number of blocks that the progress bar is divided into
when the style is \code{GTK_PROGRESS_DISCRETE}.
\strong{WARNING: \code{gtk_progress_bar_set_discrete_blocks} is deprecated and should not be used in newly-written code.}
}
\usage{gtkProgressBarSetDiscreteBlocks(object, blocks)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkProgressBar}}.}
\item{\verb{blocks}}{number of individual blocks making up the bar.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
