\alias{gtkProgressBarSetEllipsize}
\name{gtkProgressBarSetEllipsize}
\title{gtkProgressBarSetEllipsize}
\description{Sets the mode used to ellipsize (add an ellipsis: "...") the text 
if there is not enough space to render the entire string.}
\usage{gtkProgressBarSetEllipsize(object, mode)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkProgressBar}}]  a \verb{\link{GtkProgressBar}}}
\item{\verb{mode}}{[\code{\link{PangoEllipsizeMode}}]  a \code{\link{PangoEllipsizeMode}}}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
