\alias{gtkProgressBarSetFraction}
\name{gtkProgressBarSetFraction}
\title{gtkProgressBarSetFraction}
\description{Causes the progress bar to "fill in" the given fraction
of the bar. The fraction should be between 0.0 and 1.0,
inclusive.}
\usage{gtkProgressBarSetFraction(object, fraction)}
\arguments{
\item{\code{object}}{[\code{\link{GtkProgressBar}}]  a \code{\link{GtkProgressBar}}}
\item{\code{fraction}}{[numeric]  fraction of the task that's been completed}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
