\alias{gtkProgressBarSetPulseStep}
\name{gtkProgressBarSetPulseStep}
\title{gtkProgressBarSetPulseStep}
\description{Sets the fraction of total progress bar length to move the
bouncing block for each call to \code{\link{gtkProgressBarPulse}}.}
\usage{gtkProgressBarSetPulseStep(object, fraction)}
\arguments{
\item{\code{object}}{[\code{\link{GtkProgressBar}}]  a \code{\link{GtkProgressBar}}}
\item{\code{fraction}}{[numeric]  fraction between 0.0 and 1.0}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
