\alias{gtkProgressBarSetText}
\name{gtkProgressBarSetText}
\title{gtkProgressBarSetText}
\description{Causes the given \code{text} to appear superimposed on the progress bar.}
\usage{gtkProgressBarSetText(object, text)}
\arguments{
\item{\code{object}}{[\code{\link{GtkProgressBar}}]  a \code{\link{GtkProgressBar}}}
\item{\code{text}}{[character]  a UTF-8 string}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
