\alias{gtkProgressBarUpdate}
\name{gtkProgressBarUpdate}
\title{gtkProgressBarUpdate}
\description{
This function is deprecated.  Please use \code{\link{gtkProgressSetValue}} or
\code{\link{gtkProgressSetPercentage}} instead.
\strong{WARNING: \code{gtk_progress_bar_update} is deprecated and should not be used in newly-written code.}
}
\usage{gtkProgressBarUpdate(object, percentage)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkProgressBar}}.}
\item{\verb{percentage}}{the new percent complete value.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
