\alias{gtkProgressGetCurrentPercentage}
\name{gtkProgressGetCurrentPercentage}
\title{gtkProgressGetCurrentPercentage}
\description{
Returns the current progress as a percentage.
\strong{WARNING: \code{gtk_progress_get_current_percentage} is deprecated and should not be used in newly-written code.}
}
\usage{gtkProgressGetCurrentPercentage(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkProgress}}.}}
\value{[numeric] a number between 0.0 and 1.0 indicating the percentage complete.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
