\alias{gtkProgressGetValue}
\name{gtkProgressGetValue}
\title{gtkProgressGetValue}
\description{
Returns the current progress complete value.
\strong{WARNING: \code{gtk_progress_get_value} is deprecated and should not be used in newly-written code.}
}
\usage{gtkProgressGetValue(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkProgress}}.}}
\value{[numeric] the current progress complete value.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
