\alias{gtkProgressSetShowText}
\name{gtkProgressSetShowText}
\title{gtkProgressSetShowText}
\description{
Controls whether progress text is shown.
\strong{WARNING: \code{gtk_progress_set_show_text} is deprecated and should not be used in newly-written code.}
}
\usage{gtkProgressSetShowText(object, show.text)}
\arguments{
\item{\code{object}}{[\code{\link{GtkProgress}}] a \code{\link{GtkProgress}}.}
\item{\code{show.text}}{[logical] a boolean indicating whether the progress text
is shown.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
