\alias{gtkProgressSetTextAlignment}
\name{gtkProgressSetTextAlignment}
\title{gtkProgressSetTextAlignment}
\description{
Controls the alignment of the text within the progress bar area.
\strong{WARNING: \code{gtk_progress_set_text_alignment} is deprecated and should not be used in newly-written code.}
}
\usage{gtkProgressSetTextAlignment(object, x.align, y.align)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkProgress}}] a \verb{\link{GtkProgress}}.}
\item{\verb{x.align}}{[numeric] a number between 0.0 and 1.0 indicating the horizontal
alignment of the progress text within the \verb{\link{GtkProgress}}.}
\item{\verb{y.align}}{[numeric] a number between 0.0 and 1.0 indicating the vertical
alignment of the progress text within the \verb{\link{GtkProgress}}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
