\alias{gtkQuitRemove}
\name{gtkQuitRemove}
\title{gtkQuitRemove}
\description{Removes a quit handler by its identifier.}
\usage{gtkQuitRemove(quit.handler.id)}
\arguments{\item{\code{quit.handler.id}}{[numeric] Identifier for the handler returned when installing it.}}
\note{
This function is deprecated. You should use \code{GDK_PRIORITY_REDRAW} instead.
This function is deprecated. You should use \code{G_PRIORITY_HIGH} instead.
This function is deprecated. You should use \code{G_PRIORITY_DEFAULT_IDLE} instead.
This function is deprecated. You should use \code{G_PRIORITY_LOW} instead.
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
