\alias{gtkQuitRemoveByData}
\name{gtkQuitRemoveByData}
\title{gtkQuitRemoveByData}
\description{Removes a quit handler identified by its \code{data} field.}
\usage{gtkQuitRemoveByData(data)}
\arguments{\item{\code{data}}{[R object] The pointer passed as \code{data} to \code{\link{gtkQuitAdd}} or \code{\link{gtkQuitAddFull}}.}}
\note{
This function is deprecated. You should use \code{GDK_PRIORITY_REDRAW} instead.
This function is deprecated. You should use \code{G_PRIORITY_HIGH} instead.
This function is deprecated. You should use \code{G_PRIORITY_DEFAULT_IDLE} instead.
This function is deprecated. You should use \code{G_PRIORITY_LOW} instead.
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
