\alias{gtkRadioActionGetCurrentValue}
\name{gtkRadioActionGetCurrentValue}
\title{gtkRadioActionGetCurrentValue}
\description{Obtains the value property of the currently active member of 
the group to which \code{action} belongs.}
\usage{gtkRadioActionGetCurrentValue(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRadioAction}}}}
\details{Since 2.4}
\value{[integer] The value of the currently active group member}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
