\alias{gtkRadioButtonNew}
\name{gtkRadioButtonNew}
\title{gtkRadioButtonNew}
\description{Creates a new \code{\link{GtkRadioButton}}. To be of any practical value, a widget should
then be packed into the radio button.}
\usage{gtkRadioButtonNew(group = NULL, show = TRUE)}
\arguments{\item{\verb{group}}{an existing radio button group, or \code{NULL} if you are creating a new group.}}
\value{[\code{\link{GtkWidget}}] a new radio button.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
