\alias{gtkRadioButtonNewFromWidget}
\name{gtkRadioButtonNewFromWidget}
\title{gtkRadioButtonNewFromWidget}
\description{Creates a new \verb{\link{GtkRadioButton}}, adding it to the same group as \code{group}. As
with \code{\link{gtkRadioButtonNew}}, a widget should be packed into the radio button.}
\usage{gtkRadioButtonNewFromWidget(group = NULL, show = TRUE)}
\arguments{\item{\verb{group}}{[\verb{\link{GtkRadioButton}}] }}
\value{[\verb{\link{GtkWidget}}] a new radio button.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
