\alias{gtkRadioButtonNewWithLabel}
\name{gtkRadioButtonNewWithLabel}
\title{gtkRadioButtonNewWithLabel}
\description{Creates a new \verb{\link{GtkRadioButton}} with a text label.}
\usage{gtkRadioButtonNewWithLabel(group = NULL, label, show = TRUE)}
\arguments{
\item{\verb{group}}{[list] an existing radio button group, or \code{NULL} if you are creating a new
group.}
\item{\verb{label}}{[character] the text label to display next to the radio button.}
}
\value{[\verb{\link{GtkWidget}}] a new radio button.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
