\alias{gtkRadioMenuItemGetGroup}
\name{gtkRadioMenuItemGetGroup}
\title{gtkRadioMenuItemGetGroup}
\description{Returns the group to which the radio menu item belongs, as a \verb{list} of
\verb{\link{GtkRadioMenuItem}}.}
\usage{gtkRadioMenuItemGetGroup(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkRadioMenuItem}}] a \verb{\link{GtkRadioMenuItem}}.}}
\value{[list] the group of \code{radio.menu.item}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
