\alias{gtkRadioMenuItemNew}
\name{gtkRadioMenuItemNew}
\title{gtkRadioMenuItemNew}
\description{Creates a new \verb{\link{GtkRadioMenuItem}}.}
\usage{gtkRadioMenuItemNew(group = NULL, show = TRUE)}
\arguments{\item{\verb{group}}{[list] the group to which the radio menu item is to be attached}}
\value{[\verb{\link{GtkWidget}}] a new \verb{\link{GtkRadioMenuItem}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
