\alias{gtkRadioMenuItemNewFromWidget}
\name{gtkRadioMenuItemNewFromWidget}
\title{gtkRadioMenuItemNewFromWidget}
\description{Creates a new \verb{\link{GtkRadioMenuItem}} adding it to the same group as \code{group}.}
\usage{gtkRadioMenuItemNewFromWidget(group = NULL, show = TRUE)}
\arguments{\item{\verb{group}}{[\verb{\link{GtkRadioMenuItem}}]  An existing \verb{\link{GtkRadioMenuItem}}}}
\details{  Since  2.4}
\value{[\verb{\link{GtkWidget}}]  The new \verb{\link{GtkRadioMenuItem}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
