\alias{gtkRadioMenuItemNewWithLabel}
\name{gtkRadioMenuItemNewWithLabel}
\title{gtkRadioMenuItemNewWithLabel}
\description{Creates a new \code{\link{GtkRadioMenuItem}} whose child is a simple \code{\link{GtkLabel}}.}
\usage{gtkRadioMenuItemNewWithLabel(group, label, show = TRUE)}
\arguments{
\item{\code{group}}{[list] the group to which the radio menu item is to be attached}
\item{\code{label}}{[character] the text for the label}
}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkRadioMenuItem}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
