\alias{gtkRadioMenuItemNewWithLabelFromWidget}
\name{gtkRadioMenuItemNewWithLabelFromWidget}
\title{gtkRadioMenuItemNewWithLabelFromWidget}
\description{Creates a new GtkRadioMenuItem whose child is a simple GtkLabel.
The new \verb{\link{GtkRadioMenuItem}} is added to the same group as \code{group}.}
\usage{gtkRadioMenuItemNewWithLabelFromWidget(group = NULL, label, show = TRUE)}
\arguments{
\item{\verb{group}}{[\verb{\link{GtkRadioMenuItem}}]  an existing \verb{\link{GtkRadioMenuItem}} }
\item{\verb{label}}{[character]  the text for the label}
}
\details{  Since  2.4}
\value{[\verb{\link{GtkWidget}}]  The new \verb{\link{GtkRadioMenuItem}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
