\alias{gtkRadioMenuItemSetGroup}
\name{gtkRadioMenuItemSetGroup}
\title{gtkRadioMenuItemSetGroup}
\description{Sets the group of a radio menu item, or changes it.}
\usage{gtkRadioMenuItemSetGroup(object, group)}
\arguments{
\item{\code{object}}{[\code{\link{GtkRadioMenuItem}}] a \code{\link{GtkRadioMenuItem}}.}
\item{\code{group}}{[list] the new group.}
}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
