\alias{gtkRadioToolButtonNew}
\name{gtkRadioToolButtonNew}
\title{gtkRadioToolButtonNew}
\description{Creates a new \verb{\link{GtkRadioToolButton}}, adding it to \code{group}.}
\usage{gtkRadioToolButtonNew(group = NULL, show = TRUE)}
\arguments{\item{\verb{group}}{[list]  An existing radio button group, or \code{NULL} if you are creating a new group}}
\details{  Since  2.4}
\value{[\verb{\link{GtkToolItem}}]  The new \verb{\link{GtkRadioToolButton}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
