\alias{gtkRadioToolButtonNewFromWidget}
\name{gtkRadioToolButtonNewFromWidget}
\title{gtkRadioToolButtonNewFromWidget}
\description{Creates a new \verb{\link{GtkRadioToolButton}} adding it to the same group as \code{gruup}}
\usage{gtkRadioToolButtonNewFromWidget(group = NULL, show = TRUE)}
\arguments{\item{\verb{group}}{[\verb{\link{GtkRadioToolButton}}]  An existing \verb{\link{GtkRadioToolButton}}}}
\details{  Since  2.4}
\value{[\verb{\link{GtkToolItem}}]  The new \verb{\link{GtkRadioToolButton}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
