\alias{gtkRadioToolButtonNewWithStockFromWidget}
\name{gtkRadioToolButtonNewWithStockFromWidget}
\title{gtkRadioToolButtonNewWithStockFromWidget}
\description{Creates a new \verb{\link{GtkRadioToolButton}} adding it to the same group as \code{group}.
The new \verb{\link{GtkRadioToolButton}} will contain an icon and label from the
stock item indicated by \code{stock.id}.}
\usage{gtkRadioToolButtonNewWithStockFromWidget(group = NULL, stock.id, show = TRUE)}
\arguments{
\item{\verb{group}}{[\verb{\link{GtkRadioToolButton}}]  An existing \verb{\link{GtkRadioToolButton}}.}
\item{\verb{stock.id}}{[character]  the name of a stock item }
}
\details{  Since  2.4}
\value{[\verb{\link{GtkToolItem}}]  A new \verb{\link{GtkRadioToolButton}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
