\alias{gtkRadioToolButtonSetGroup}
\name{gtkRadioToolButtonSetGroup}
\title{gtkRadioToolButtonSetGroup}
\description{Adds \code{button} to \code{group}, removing it from the group it belonged to before.}
\usage{gtkRadioToolButtonSetGroup(object, group)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkRadioToolButton}}]  a \verb{\link{GtkRadioToolButton}}}
\item{\verb{group}}{[list]  an existing radio button group}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
