\alias{gtkRangeGetAdjustment}
\name{gtkRangeGetAdjustment}
\title{gtkRangeGetAdjustment}
\description{Get the \verb{\link{GtkAdjustment}} which is the "model" object for \verb{\link{GtkRange}}.
See \code{\link{gtkRangeSetAdjustment}} for details.
The return value does not have a reference added, so should not
be unreferenced.}
\usage{gtkRangeGetAdjustment(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkRange}}]  a \verb{\link{GtkRange}}}}
\value{[\verb{\link{GtkAdjustment}}]  a \verb{\link{GtkAdjustment}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
