\alias{gtkRangeGetFlippable}
\name{gtkRangeGetFlippable}
\title{gtkRangeGetFlippable}
\description{Gets the value set by \code{\link{gtkRangeSetFlippable}}.}
\usage{gtkRangeGetFlippable(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRange}}}}
\details{Since 2.18}
\value{[logical] \code{TRUE} if the range is flippable}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
