\alias{gtkRangeGetMinSliderSize}
\name{gtkRangeGetMinSliderSize}
\title{gtkRangeGetMinSliderSize}
\description{This function is useful mainly for \code{\link{GtkRange}} subclasses.}
\usage{gtkRangeGetMinSliderSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRange}}}}
\details{See \code{\link{gtkRangeSetMinSliderSize}}.
  Since 2.20}
\value{[integer] The minimum size of the range's slider.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
