\alias{gtkRangeGetRestrictToFillLevel}
\name{gtkRangeGetRestrictToFillLevel}
\title{gtkRangeGetRestrictToFillLevel}
\description{Gets whether the range is restricted to the fill level.}
\usage{gtkRangeGetRestrictToFillLevel(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkRange}}}}
\details{Since 2.12}
\value{[logical] \code{TRUE} if \code{range} is restricted to the fill level.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
