\alias{gtkRangeGetSliderSizeFixed}
\name{gtkRangeGetSliderSizeFixed}
\title{gtkRangeGetSliderSizeFixed}
\description{This function is useful mainly for \code{\link{GtkRange}} subclasses.}
\usage{gtkRangeGetSliderSizeFixed(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRange}}}}
\details{See \code{\link{gtkRangeSetSliderSizeFixed}}.
  Since 2.20}
\value{[logical] whether the range's slider has a fixed size.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
