\alias{gtkRangeGetUpperStepperSensitivity}
\name{gtkRangeGetUpperStepperSensitivity}
\title{gtkRangeGetUpperStepperSensitivity}
\description{Gets the sensitivity policy for the stepper that points to the
'upper' end of the GtkRange's adjustment.}
\usage{gtkRangeGetUpperStepperSensitivity(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkRange}}]  a \verb{\link{GtkRange}}}}
\details{  Since  2.10}
\value{[\code{\link{GtkSensitivityType}}]  The upper stepper's sensitivity policy.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
