\alias{gtkRangeGetValue}
\name{gtkRangeGetValue}
\title{gtkRangeGetValue}
\description{Gets the current value of the range.}
\usage{gtkRangeGetValue(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkRange}}]  a \verb{\link{GtkRange}}}}
\value{[numeric]  current value of the range.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
