\alias{gtkRangeSetFlippable}
\name{gtkRangeSetFlippable}
\title{gtkRangeSetFlippable}
\description{If a range is flippable, it will switch its direction if it is
horizontal and its direction is \code{GTK_TEXT_DIR_RTL}.}
\usage{gtkRangeSetFlippable(object, flippable)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRange}}}
\item{\verb{flippable}}{\code{TRUE} to make the range flippable}
}
\details{See \code{\link{gtkWidgetGetDirection}}.
  Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
