\alias{gtkRangeSetIncrements}
\name{gtkRangeSetIncrements}
\title{gtkRangeSetIncrements}
\description{Sets the step and page sizes for the range.
The step size is used when the user clicks the \code{\link{GtkScrollbar}}
arrows or moves \code{\link{GtkScale}} via arrow keys. The page size
is used for example when moving via Page Up or Page Down keys.}
\usage{gtkRangeSetIncrements(object, step, page)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRange}}}
\item{\verb{step}}{step size}
\item{\verb{page}}{page size}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
