\alias{gtkRangeSetMinSliderSize}
\name{gtkRangeSetMinSliderSize}
\title{gtkRangeSetMinSliderSize}
\description{Sets the minimum size of the range's slider.}
\usage{gtkRangeSetMinSliderSize(object, min.size)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRange}}}
\item{\verb{min.size}}{The slider's minimum size}
}
\details{This function is useful mainly for \code{\link{GtkRange}} subclasses.
  Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
