\alias{gtkRangeSetRestrictToFillLevel}
\name{gtkRangeSetRestrictToFillLevel}
\title{gtkRangeSetRestrictToFillLevel}
\description{Sets whether the slider is restricted to the fill level. See
\code{\link{gtkRangeSetFillLevel}} for a general description of the fill
level concept.}
\usage{gtkRangeSetRestrictToFillLevel(object, restrict.to.fill.level)}
\arguments{
\item{\code{object}}{[\code{\link{GtkRange}}]                   A \code{\link{GtkRange}}}
\item{\code{restrict.to.fill.level}}{[logical]  Whether the fill level restricts slider movement.}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
