\alias{gtkRangeSetShowFillLevel}
\name{gtkRangeSetShowFillLevel}
\title{gtkRangeSetShowFillLevel}
\description{Sets whether a graphical fill level is show on the trough. See
\code{\link{gtkRangeSetFillLevel}} for a general description of the fill
level concept.}
\usage{gtkRangeSetShowFillLevel(object, show.fill.level)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkRange}}}
\item{\verb{show.fill.level}}{Whether a fill level indicator graphics is shown.}
}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
