\alias{gtkRangeSetSliderSizeFixed}
\name{gtkRangeSetSliderSizeFixed}
\title{gtkRangeSetSliderSizeFixed}
\description{Sets whether the range's slider has a fixed size, or a size that
depends on it's adjustment's page size.}
\usage{gtkRangeSetSliderSizeFixed(object, size.fixed)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRange}}}
\item{\verb{size.fixed}}{\code{TRUE} to make the slider size constant}
}
\details{This function is useful mainly for \code{\link{GtkRange}} subclasses.
  Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
