\alias{gtkRcAddDefaultFile}
\name{gtkRcAddDefaultFile}
\title{gtkRcAddDefaultFile}
\description{Adds a file to the list of files to be parsed at the
end of \code{\link{gtkInit}}.}
\usage{gtkRcAddDefaultFile(filename)}
\arguments{\item{\verb{filename}}{the pathname to the file. If \code{filename} is not absolute, it
   is searched in the current directory.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
