\alias{gtkRcFindModuleInPath}
\name{gtkRcFindModuleInPath}
\title{gtkRcFindModuleInPath}
\description{Searches for a theme engine in the GTK+ search path. This function
is not useful for applications and should not be used.}
\usage{gtkRcFindModuleInPath(module.file)}
\arguments{\item{\verb{module.file}}{name of a theme engine}}
\value{[character] The filename,
otherwise \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
