\alias{gtkRcGetDefaultFiles}
\name{gtkRcGetDefaultFiles}
\title{gtkRcGetDefaultFiles}
\description{Retrieves the current list of RC files that will be parsed
at the end of \code{\link{gtkInit}}.}
\usage{gtkRcGetDefaultFiles()}
\value{[character] A list of filenames.
If you want to store this information, you should make a copy.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
