\alias{gtkRcGetImModuleFile}
\name{gtkRcGetImModuleFile}
\title{gtkRcGetImModuleFile}
\description{Obtains the path to the IM modules file. See the documentation
of the \env{GTK_IM_MODULE_FILE}
environment variable for more details.}
\usage{gtkRcGetImModuleFile()}
\value{[character] a newly-allocated string containing the name of the file
listing the IM modules available for loading}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
