\alias{gtkRcParse}
\name{gtkRcParse}
\title{gtkRcParse}
\description{Parses a given resource file.}
\usage{gtkRcParse(filename)}
\arguments{\item{\verb{filename}}{the filename of a file to parse. If \code{filename} is not absolute, it
  is searched in the current directory.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
