\alias{gtkRcParseString}
\name{gtkRcParseString}
\title{gtkRcParseString}
\description{Parses resource information directly from a string.}
\usage{gtkRcParseString(rc.string)}
\arguments{\item{\code{rc.string}}{[character] a string to parse.}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
