\alias{gtkRcReparseAll}
\name{gtkRcReparseAll}
\title{gtkRcReparseAll}
\description{If the modification time on any previously read file for the
default \code{\link{GtkSettings}} has changed, discard all style information
and then reread all previously read RC files.}
\usage{gtkRcReparseAll()}
\value{[logical] \code{TRUE} if the files were reread.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
