\alias{gtkRcSetDefaultFiles}
\name{gtkRcSetDefaultFiles}
\title{gtkRcSetDefaultFiles}
\description{Sets the list of files that GTK+ will read at the
end of \code{\link{gtkInit}}.}
\usage{gtkRcSetDefaultFiles(filenames)}
\arguments{\item{\verb{filenames}}{A list of filenames.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
