\alias{gtkRcStyleCopy}
\name{gtkRcStyleCopy}
\title{gtkRcStyleCopy}
\description{Makes a copy of the specified \verb{\link{GtkRcStyle}}. This function
will correctly copy an RC style that is a member of a class
derived from \verb{\link{GtkRcStyle}}.}
\usage{gtkRcStyleCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkRcStyle}}]  the style to copy}}
\value{[\verb{\link{GtkRcStyle}}]  the resulting \verb{\link{GtkRcStyle}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
