\alias{gtkRcStyleNew}
\name{gtkRcStyleNew}
\title{gtkRcStyleNew}
\description{Creates a new \code{\link{GtkRcStyle}} with no fields set and
a reference count of 1.}
\usage{gtkRcStyleNew()}
\value{[\code{\link{GtkRcStyle}}] the newly-created \code{\link{GtkRcStyle}}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
