\alias{gtkRecentActionSetShowNumbers}
\name{gtkRecentActionSetShowNumbers}
\title{gtkRecentActionSetShowNumbers}
\description{Sets whether a number should be added to the items shown by the
widgets representing \code{action}. The numbers are shown to provide
a unique character for a mnemonic to be used inside the menu item's
label. Only the first ten items get a number to avoid clashes.}
\usage{gtkRecentActionSetShowNumbers(object, show.numbers)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentAction}}}
\item{\verb{show.numbers}}{\code{TRUE} if the shown items should be numbered}
}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
