\alias{gtkRecentChooserGetCurrentItem}
\name{gtkRecentChooserGetCurrentItem}
\title{gtkRecentChooserGetCurrentItem}
\description{Gets the \verb{\link{GtkRecentInfo}} currently selected by \code{chooser}.}
\usage{gtkRecentChooserGetCurrentItem(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkRecentChooser}}]  a \verb{\link{GtkRecentChooser}}}}
\details{  Since  2.10}
\value{[\verb{\link{GtkRecentInfo}}]  a \verb{\link{GtkRecentInfo}}.  Use \code{\link{gtkRecentInfoUnref}} when
  when you have finished using it.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
