\alias{gtkRecentChooserGetCurrentUri}
\name{gtkRecentChooserGetCurrentUri}
\title{gtkRecentChooserGetCurrentUri}
\description{Gets the URI currently selected by \code{chooser}.}
\usage{gtkRecentChooserGetCurrentUri(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}}
\details{Since 2.10}
\value{[character] a newly allocated string holding a URI.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
