\alias{gtkRecentChooserGetFilter}
\name{gtkRecentChooserGetFilter}
\title{gtkRecentChooserGetFilter}
\description{Gets the \verb{\link{GtkRecentFilter}} object currently used by \code{chooser} to affect
the display of the recently used resources.}
\usage{gtkRecentChooserGetFilter(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkRecentChooser}}]  a \verb{\link{GtkRecentChooser}}}}
\details{  Since  2.10}
\value{[\verb{\link{GtkRecentFilter}}]  a \verb{\link{GtkRecentFilter}} object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
