\alias{gtkRecentChooserGetLimit}
\name{gtkRecentChooserGetLimit}
\title{gtkRecentChooserGetLimit}
\description{Gets the number of items returned by \code{\link{gtkRecentChooserGetItems}}
and \code{\link{gtkRecentChooserGetUris}}.}
\usage{gtkRecentChooserGetLimit(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}}
\details{Since 2.10}
\value{[integer] A positive integer, or -1 meaning that all items are
returned.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
