\alias{gtkRecentChooserGetLocalOnly}
\name{gtkRecentChooserGetLocalOnly}
\title{gtkRecentChooserGetLocalOnly}
\description{Gets whether only local resources should be shown in the recently used
resources selector.  See \code{\link{gtkRecentChooserSetLocalOnly}}}
\usage{gtkRecentChooserGetLocalOnly(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}}
\details{Since 2.10}
\value{[logical] \code{TRUE} if only local resources should be shown.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
