\alias{gtkRecentChooserGetSelectMultiple}
\name{gtkRecentChooserGetSelectMultiple}
\title{gtkRecentChooserGetSelectMultiple}
\description{Gets whether \code{chooser} can select multiple items.}
\usage{gtkRecentChooserGetSelectMultiple(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkRecentChooser}}]  a \code{\link{GtkRecentChooser}}}}
\details{  Since  2.10}
\value{[logical]  \code{TRUE} if \code{chooser} can select more than one item.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
