\alias{gtkRecentChooserGetShowIcons}
\name{gtkRecentChooserGetShowIcons}
\title{gtkRecentChooserGetShowIcons}
\description{Retrieves whether \code{chooser} should show an icon near the resource.}
\usage{gtkRecentChooserGetShowIcons(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}}
\details{Since 2.10}
\value{[logical] \code{TRUE} if the icons should be displayed, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
