\alias{gtkRecentChooserGetShowNotFound}
\name{gtkRecentChooserGetShowNotFound}
\title{gtkRecentChooserGetShowNotFound}
\description{Retrieves whether \code{chooser} should show the recently used resources that
were not found.}
\usage{gtkRecentChooserGetShowNotFound(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}}
\details{Since 2.10}
\value{[logical] \code{TRUE} if the resources not found should be displayed, and
\code{FALSE} otheriwse.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
