\alias{gtkRecentChooserGetShowPrivate}
\name{gtkRecentChooserGetShowPrivate}
\title{gtkRecentChooserGetShowPrivate}
\description{Returns whether \code{chooser} should display recently used resources
registered as private.}
\usage{gtkRecentChooserGetShowPrivate(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}}
\details{Since 2.10}
\value{[logical] \code{TRUE} if the recent chooser should show private items,
\code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
