\alias{gtkRecentChooserGetShowTips}
\name{gtkRecentChooserGetShowTips}
\title{gtkRecentChooserGetShowTips}
\description{Gets whether \code{chooser} should display tooltips containing the full path
of a recently user resource.}
\usage{gtkRecentChooserGetShowTips(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}}
\details{Since 2.10}
\value{[logical] \code{TRUE} if the recent chooser should show tooltips,
\code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
