\alias{gtkRecentChooserListFilters}
\name{gtkRecentChooserListFilters}
\title{gtkRecentChooserListFilters}
\description{Gets the \verb{\link{GtkRecentFilter}} objects held by \code{chooser}.}
\usage{gtkRecentChooserListFilters(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkRecentChooser}}]  a \verb{\link{GtkRecentChooser}}}}
\details{  Since  2.10}
\value{[list]  A singly linked list of \verb{\link{GtkRecentFilter}} objects.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
