\alias{gtkRecentChooserMenuGetShowNumbers}
\name{gtkRecentChooserMenuGetShowNumbers}
\title{gtkRecentChooserMenuGetShowNumbers}
\description{Returns the value set by \code{\link{gtkRecentChooserMenuSetShowNumbers}}.}
\usage{gtkRecentChooserMenuGetShowNumbers(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentChooserMenu}}}}
\details{Since 2.10}
\value{[logical] \code{TRUE} if numbers should be shown.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
