\alias{gtkRecentChooserMenuSetShowNumbers}
\name{gtkRecentChooserMenuSetShowNumbers}
\title{gtkRecentChooserMenuSetShowNumbers}
\description{Sets whether a number should be added to the items of \code{menu}.  The
numbers are shown to provide a unique character for a mnemonic to
be used inside the menu item's label.  Only the first the items
get a number to avoid clashes.}
\usage{gtkRecentChooserMenuSetShowNumbers(object, show.numbers)}
\arguments{
\item{\code{object}}{[\code{\link{GtkRecentChooserMenu}}]  a \code{\link{GtkRecentChooserMenu}}}
\item{\code{show.numbers}}{[logical]  whether to show numbers}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
