\alias{gtkRecentChooserSelectAll}
\name{gtkRecentChooserSelectAll}
\title{gtkRecentChooserSelectAll}
\description{Selects all the items inside \code{chooser}, if the \code{chooser} supports
multiple selection.}
\usage{gtkRecentChooserSelectAll(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
