\alias{gtkRecentChooserSetFilter}
\name{gtkRecentChooserSetFilter}
\title{gtkRecentChooserSetFilter}
\description{Sets \code{filter} as the current \code{\link{GtkRecentFilter}} object used by \code{chooser}
to affect the displayed recently used resources.}
\usage{gtkRecentChooserSetFilter(object, filter)}
\arguments{
\item{\code{object}}{[\code{\link{GtkRecentChooser}}]  a \code{\link{GtkRecentChooser}}}
\item{\code{filter}}{[\code{\link{GtkRecentFilter}}]  a \code{\link{GtkRecentFilter}}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
