\alias{gtkRecentChooserSetLimit}
\name{gtkRecentChooserSetLimit}
\title{gtkRecentChooserSetLimit}
\description{Sets the number of items that should be returned by
\code{\link{gtkRecentChooserGetItems}} and \code{\link{gtkRecentChooserGetUris}}.}
\usage{gtkRecentChooserSetLimit(object, limit)}
\arguments{
\item{\code{object}}{[\code{\link{GtkRecentChooser}}]  a \code{\link{GtkRecentChooser}}}
\item{\code{limit}}{[integer]  a positive integer, or -1 for all items}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
