\alias{gtkRecentChooserSetSelectMultiple}
\name{gtkRecentChooserSetSelectMultiple}
\title{gtkRecentChooserSetSelectMultiple}
\description{Sets whether \code{chooser} can select multiple items.}
\usage{gtkRecentChooserSetSelectMultiple(object, select.multiple)}
\arguments{
\item{\code{object}}{[\code{\link{GtkRecentChooser}}]  a \code{\link{GtkRecentChooser}}}
\item{\code{select.multiple}}{[logical] \code{TRUE} if \code{chooser} can select more than one item}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
