\alias{gtkRecentChooserSetShowIcons}
\name{gtkRecentChooserSetShowIcons}
\title{gtkRecentChooserSetShowIcons}
\description{Sets whether \code{chooser} should show an icon near the resource when
displaying it.}
\usage{gtkRecentChooserSetShowIcons(object, show.icons)}
\arguments{
\item{\code{object}}{[\code{\link{GtkRecentChooser}}]  a \code{\link{GtkRecentChooser}}}
\item{\code{show.icons}}{[logical]  whether to show an icon near the resource}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
