\alias{gtkRecentChooserSetShowNotFound}
\name{gtkRecentChooserSetShowNotFound}
\title{gtkRecentChooserSetShowNotFound}
\description{Sets whether \code{chooser} should display the recently used resources that
it didn't find.  This only applies to local resources.}
\usage{gtkRecentChooserSetShowNotFound(object, show.not.found)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}
\item{\verb{show.not.found}}{whether to show the local items we didn't find}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
