\alias{gtkRecentChooserSetShowPrivate}
\name{gtkRecentChooserSetShowPrivate}
\title{gtkRecentChooserSetShowPrivate}
\description{Whether to show recently used resources marked registered as private.}
\usage{gtkRecentChooserSetShowPrivate(object, show.private)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}
\item{\verb{show.private}}{\code{TRUE} to show private items, \code{FALSE} otherwise}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
