\alias{gtkRecentChooserUnselectAll}
\name{gtkRecentChooserUnselectAll}
\title{gtkRecentChooserUnselectAll}
\description{Unselects all the items inside \code{chooser}.}
\usage{gtkRecentChooserUnselectAll(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkRecentChooser}}]  a \code{\link{GtkRecentChooser}}}}
\details{  Since  2.10}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
