\alias{gtkRecentChooserUnselectUri}
\name{gtkRecentChooserUnselectUri}
\title{gtkRecentChooserUnselectUri}
\description{Unselects \code{uri} inside \code{chooser}.}
\usage{gtkRecentChooserUnselectUri(object, uri)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkRecentChooser}}]  a \verb{\link{GtkRecentChooser}}}
\item{\verb{uri}}{[character]  a URI}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
