\alias{gtkRecentFilterAddGroup}
\name{gtkRecentFilterAddGroup}
\title{gtkRecentFilterAddGroup}
\description{Adds a rule that allows resources based on the name of the group
to which they belong}
\usage{gtkRecentFilterAddGroup(object, group)}
\arguments{
\item{\code{object}}{[\code{\link{GtkRecentFilter}}]  a \code{\link{GtkRecentFilter}}}
\item{\code{group}}{[character]  a group name}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
