\alias{gtkRecentFilterAddMimeType}
\name{gtkRecentFilterAddMimeType}
\title{gtkRecentFilterAddMimeType}
\description{Adds a rule that allows resources based on their registered MIME type.}
\usage{gtkRecentFilterAddMimeType(object, mime.type)}
\arguments{
\item{\code{object}}{[\code{\link{GtkRecentFilter}}]  a \code{\link{GtkRecentFilter}}}
\item{\code{mime.type}}{[character]  a MIME type}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
