\alias{gtkRecentFilterAddPattern}
\name{gtkRecentFilterAddPattern}
\title{gtkRecentFilterAddPattern}
\description{Adds a rule that allows resources based on a pattern matching their
display name.}
\usage{gtkRecentFilterAddPattern(object, pattern)}
\arguments{
\item{\code{object}}{[\code{\link{GtkRecentFilter}}]  a \code{\link{GtkRecentFilter}}}
\item{\code{pattern}}{[character]  a file pattern}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
