\alias{gtkRecentFilterAddPixbufFormats}
\name{gtkRecentFilterAddPixbufFormats}
\title{gtkRecentFilterAddPixbufFormats}
\description{Adds a rule allowing image files in the formats supported
by GdkPixbuf.}
\usage{gtkRecentFilterAddPixbufFormats(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentFilter}}}}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
