\alias{gtkRecentFilterSetName}
\name{gtkRecentFilterSetName}
\title{gtkRecentFilterSetName}
\description{Sets the human-readable name of the filter; this is the string
that will be displayed in the recently used resources selector
user interface if there is a selectable list of filters.}
\usage{gtkRecentFilterSetName(object, name)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentFilter}}}
\item{\verb{name}}{then human readable name of \code{filter}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
