\alias{gtkRecentInfoExists}
\name{gtkRecentInfoExists}
\title{gtkRecentInfoExists}
\description{Checks whether the resource pointed by \code{info} still exists.  At
the moment this check is done only on resources pointing to local files.}
\usage{gtkRecentInfoExists(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentInfo}}}}
\details{Since 2.10}
\value{[logical] \code{TRUE} if the resource exists}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
