\alias{gtkRecentInfoGetAdded}
\name{gtkRecentInfoGetAdded}
\title{gtkRecentInfoGetAdded}
\description{Gets the timestamp (seconds from system's Epoch) when the resource
was added to the recently used resources list.}
\usage{gtkRecentInfoGetAdded(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentInfo}}}}
\details{Since 2.10}
\value{[integer] the number of seconds elapsed from system's Epoch when
the resource was added to the list, or -1 on failure.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
