\alias{gtkRecentInfoGetAge}
\name{gtkRecentInfoGetAge}
\title{gtkRecentInfoGetAge}
\description{Gets the number of days elapsed since the last update of the resource
pointed by \code{info}.}
\usage{gtkRecentInfoGetAge(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkRecentInfo}}]  a \code{\link{GtkRecentInfo}}}}
\details{  Since  2.10}
\value{[integer]  a positive integer containing the number of days elapsed
  since the time this resource was last modified.  }
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
