\alias{gtkRecentInfoGetDescription}
\name{gtkRecentInfoGetDescription}
\title{gtkRecentInfoGetDescription}
\description{Gets the (short) description of the resource.}
\usage{gtkRecentInfoGetDescription(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkRecentInfo}}]  a \verb{\link{GtkRecentInfo}}}}
\details{  Since  2.10}
\value{[character]  the description of the resource.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
