\alias{gtkRecentInfoGetDisplayName}
\name{gtkRecentInfoGetDisplayName}
\title{gtkRecentInfoGetDisplayName}
\description{Gets the name of the resource.  If none has been defined, the basename
of the resource is obtained.}
\usage{gtkRecentInfoGetDisplayName(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkRecentInfo}}]  a \verb{\link{GtkRecentInfo}}}}
\details{  Since  2.10}
\value{[character]  the display name of the resource.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
