\alias{gtkRecentInfoGetIcon}
\name{gtkRecentInfoGetIcon}
\title{gtkRecentInfoGetIcon}
\description{Retrieves the icon of size \code{size} associated to the resource MIME type.}
\usage{gtkRecentInfoGetIcon(object, size)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentInfo}}}
\item{\verb{size}}{the size of the icon in pixels}
}
\details{Since 2.10}
\value{[\code{\link{GdkPixbuf}}] a \code{\link{GdkPixbuf}} containing the icon, or \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
