\alias{gtkRecentInfoGetUri}
\name{gtkRecentInfoGetUri}
\title{gtkRecentInfoGetUri}
\description{Gets the URI of the resource.}
\usage{gtkRecentInfoGetUri(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkRecentInfo}}]  a \code{\link{GtkRecentInfo}}}}
\details{  Since  2.10}
\value{[character]  the URI of the resource.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
