\alias{gtkRecentInfoGetUriDisplay}
\name{gtkRecentInfoGetUriDisplay}
\title{gtkRecentInfoGetUriDisplay}
\description{Gets a displayable version of the resource's URI.}
\usage{gtkRecentInfoGetUriDisplay(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkRecentInfo}}]  a \code{\link{GtkRecentInfo}}}}
\details{  Since  2.10}
\value{[character]  a UTF-8 string containing the resource's URI or \code{NULL}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
