\alias{gtkRecentInfoGetVisited}
\name{gtkRecentInfoGetVisited}
\title{gtkRecentInfoGetVisited}
\description{Gets the timestamp (seconds from system's Epoch) when the resource
was last visited.}
\usage{gtkRecentInfoGetVisited(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentInfo}}}}
\details{Since 2.10}
\value{[integer] the number of seconds elapsed from system's Epoch when
the resource was last visited, or -1 on failure.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
