\alias{gtkRecentInfoHasGroup}
\name{gtkRecentInfoHasGroup}
\title{gtkRecentInfoHasGroup}
\description{Checks whether \code{group.name} appears inside the groups registered for the
recently used item \code{info}.}
\usage{gtkRecentInfoHasGroup(object, group.name)}
\arguments{
\item{\code{object}}{[\code{\link{GtkRecentInfo}}]  a \code{\link{GtkRecentInfo}}}
\item{\code{group.name}}{[character]  name of a group}
}
\details{  Since  2.10}
\value{[logical] \code{TRUE} if the group was found.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
