\alias{gtkRecentInfoIsLocal}
\name{gtkRecentInfoIsLocal}
\title{gtkRecentInfoIsLocal}
\description{Checks whether the resource is local or not by looking at the
scheme of its URI.}
\usage{gtkRecentInfoIsLocal(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentInfo}}}}
\details{Since 2.10}
\value{[logical] \code{TRUE} if the resource is local.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
