\alias{gtkRecentInfoLastApplication}
\name{gtkRecentInfoLastApplication}
\title{gtkRecentInfoLastApplication}
\description{Gets the name of the last application that have registered the
recently used resource represented by \code{info}.}
\usage{gtkRecentInfoLastApplication(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentInfo}}}}
\details{Since 2.10}
\value{[character] an application name.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
