\alias{gtkRecentInfoMatch}
\name{gtkRecentInfoMatch}
\title{gtkRecentInfoMatch}
\description{Checks whether two \verb{\link{GtkRecentInfo}} structures point to the same
resource.}
\usage{gtkRecentInfoMatch(object, info.b)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkRecentInfo}}]  a \verb{\link{GtkRecentInfo}}}
\item{\verb{info.b}}{[\verb{\link{GtkRecentInfo}}]  a \verb{\link{GtkRecentInfo}}}
}
\details{  Since  2.10}
\value{[logical] \code{TRUE} if both \verb{\link{GtkRecentInfo}} structures point to se same
  resource, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
