\alias{gtkRecentManagerGetDefault}
\name{gtkRecentManagerGetDefault}
\title{gtkRecentManagerGetDefault}
\description{Gets a unique instance of \verb{\link{GtkRecentManager}}, that you can share
in your application without caring about memory management.}
\usage{gtkRecentManagerGetDefault()}
\details{  Since  2.10}
\value{[\verb{\link{GtkRecentManager}}]  A unique \verb{\link{GtkRecentManager}}. Do not ref or unref it.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
