\alias{gtkRecentManagerGetLimit}
\name{gtkRecentManagerGetLimit}
\title{gtkRecentManagerGetLimit}
\description{Gets the maximum number of items that the \code{\link{gtkRecentManagerGetItems}}
function should return.}
\usage{gtkRecentManagerGetLimit(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentManager}}}}
\details{Since 2.10}
\value{[integer] the number of items to return, or -1 for every item.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
