\alias{gtkRecentManagerHasItem}
\name{gtkRecentManagerHasItem}
\title{gtkRecentManagerHasItem}
\description{Checks whether there is a recently used resource registered
with \code{uri} inside the recent manager.}
\usage{gtkRecentManagerHasItem(object, uri)}
\arguments{
\item{\code{object}}{[\code{\link{GtkRecentManager}}]  a \code{\link{GtkRecentManager}}}
\item{\code{uri}}{[character]  a URI}
}
\details{  Since  2.10}
\value{[logical]  \code{TRUE} if the resource was found, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
