\alias{gtkRecentManagerSetScreen}
\name{gtkRecentManagerSetScreen}
\title{gtkRecentManagerSetScreen}
\description{Sets the screen for a recent manager; the screen is used to
track the user's currently configured recently used documents
storage.}
\usage{gtkRecentManagerSetScreen(object, screen)}
\arguments{
\item{\code{object}}{[\code{\link{GtkRecentManager}}]  a \code{\link{GtkRecentManager}}}
\item{\code{screen}}{[\code{\link{GdkScreen}}]  a \code{\link{GdkScreen}}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
