\alias{gtkRequisitionCopy}
\name{gtkRequisitionCopy}
\title{gtkRequisitionCopy}
\description{Copies a \code{\link{GtkRequisition}}.}
\usage{gtkRequisitionCopy(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkRequisition}}]  a \code{\link{GtkRequisition}}}}
\value{[\code{\link{GtkRequisition}}]  a copy of \code{requisition}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
