\alias{gtkRgbToHsv}
\name{gtkRgbToHsv}
\title{gtkRgbToHsv}
\description{Converts a color from RGB space to HSV.
Input values must be in the [0.0, 1.0] range;
output values will be in the same range.}
\usage{gtkRgbToHsv(r, g, h, s, v)}
\arguments{
\item{\verb{r}}{Red}
\item{\verb{g}}{Green}
\item{\verb{h}}{Return value for the hue component}
\item{\verb{s}}{Return value for the saturation component}
\item{\verb{v}}{Return value for the value component}
}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
