\alias{gtkScaleButtonGetAdjustment}
\name{gtkScaleButtonGetAdjustment}
\title{gtkScaleButtonGetAdjustment}
\description{Gets the \verb{\link{GtkAdjustment}} associated with the \verb{\link{GtkScaleButton}}'s scale.
See \code{\link{gtkRangeGetAdjustment}} for details.}
\usage{gtkScaleButtonGetAdjustment(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkScaleButton}}]  a \verb{\link{GtkScaleButton}}}}
\details{  Since  2.12}
\value{[\verb{\link{GtkAdjustment}}]  the adjustment associated with the scale}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
