\alias{gtkScaleButtonGetMinusButton}
\name{gtkScaleButtonGetMinusButton}
\title{gtkScaleButtonGetMinusButton}
\description{Retrieves the minus button of the \code{\link{GtkScaleButton}}.}
\usage{gtkScaleButtonGetMinusButton(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkScaleButton}}}}
\details{Since 2.14}
\value{[\code{\link{GtkWidget}}] the minus button of the \code{\link{GtkScaleButton}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
