\alias{gtkScaleButtonGetOrientation}
\name{gtkScaleButtonGetOrientation}
\title{gtkScaleButtonGetOrientation}
\description{
Gets the orientation of the \code{\link{GtkScaleButton}}'s popup window.
\strong{WARNING: \code{gtk_scale_button_get_orientation} has been deprecated since version 2.16 and should not be used in newly-written code. Use \code{\link{gtkOrientableGetOrientation}} instead.}
}
\usage{gtkScaleButtonGetOrientation(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkScaleButton}}}}
\details{Since 2.14}
\value{[\code{\link{GtkOrientation}}] the \code{\link{GtkScaleButton}}'s orientation.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
