\alias{gtkScaleButtonGetPlusButton}
\name{gtkScaleButtonGetPlusButton}
\title{gtkScaleButtonGetPlusButton}
\description{Retrieves the plus button of the \code{\link{GtkScaleButton}}.}
\usage{gtkScaleButtonGetPlusButton(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkScaleButton}}}}
\details{Since 2.14}
\value{[\code{\link{GtkWidget}}] the plus button of the \code{\link{GtkScaleButton}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
