\alias{gtkScaleButtonGetPopup}
\name{gtkScaleButtonGetPopup}
\title{gtkScaleButtonGetPopup}
\description{Retrieves the popup of the \code{\link{GtkScaleButton}}.}
\usage{gtkScaleButtonGetPopup(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkScaleButton}}}}
\details{Since 2.14}
\value{[\code{\link{GtkWidget}}] the popup of the \code{\link{GtkScaleButton}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
