\alias{gtkScaleButtonGetValue}
\name{gtkScaleButtonGetValue}
\title{gtkScaleButtonGetValue}
\description{Gets the current value of the scale button.}
\usage{gtkScaleButtonGetValue(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkScaleButton}}}}
\details{Since 2.12}
\value{[numeric] current value of the scale button}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
