\alias{gtkScaleButtonSetAdjustment}
\name{gtkScaleButtonSetAdjustment}
\title{gtkScaleButtonSetAdjustment}
\description{Sets the \code{\link{GtkAdjustment}} to be used as a model 
for the \code{\link{GtkScaleButton}}'s scale.
See \code{\link{gtkRangeSetAdjustment}} for details.}
\usage{gtkScaleButtonSetAdjustment(object, adjustment)}
\arguments{
\item{\code{object}}{[\code{\link{GtkScaleButton}}]  a \code{\link{GtkScaleButton}}}
\item{\code{adjustment}}{[\code{\link{GtkAdjustment}}]  a \code{\link{GtkAdjustment}}}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
