\alias{gtkScaleButtonSetIcons}
\name{gtkScaleButtonSetIcons}
\title{gtkScaleButtonSetIcons}
\description{Sets the icons to be used by the scale button.
For details, see the \verb{"icons"} property.}
\usage{gtkScaleButtonSetIcons(object, icons)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkScaleButton}}}
\item{\verb{icons}}{a list of icon names}
}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
