\alias{gtkScaleButtonSetOrientation}
\name{gtkScaleButtonSetOrientation}
\title{gtkScaleButtonSetOrientation}
\description{
Sets the orientation of the \code{\link{GtkScaleButton}}'s popup window.
\strong{WARNING: \code{gtk_scale_button_set_orientation} has been deprecated since version 2.16 and should not be used in newly-written code. Use \code{\link{gtkOrientableSetOrientation}} instead.}
}
\usage{gtkScaleButtonSetOrientation(object, orientation)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkScaleButton}}}
\item{\verb{orientation}}{the new orientation}
}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
