\alias{gtkScaleClearMarks}
\name{gtkScaleClearMarks}
\title{gtkScaleClearMarks}
\description{Removes any marks that have been added with \code{\link{gtkScaleAddMark}}.}
\usage{gtkScaleClearMarks(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkScale}}}}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
