\alias{gtkScaleGetDigits}
\name{gtkScaleGetDigits}
\title{gtkScaleGetDigits}
\description{Gets the number of decimal places that are displayed in the value.}
\usage{gtkScaleGetDigits(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkScale}}] a \code{\link{GtkScale}}.}}
\value{[integer] the number of decimal places that are displayed.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
