\alias{gtkScaleGetDrawValue}
\name{gtkScaleGetDrawValue}
\title{gtkScaleGetDrawValue}
\description{Returns whether the current value is displayed as a string 
next to the slider.}
\usage{gtkScaleGetDrawValue(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkScale}}}}
\value{[logical] whether the current value is displayed as a string}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
