\alias{gtkScaleGetLayout}
\name{gtkScaleGetLayout}
\title{gtkScaleGetLayout}
\description{Gets the \verb{\link{PangoLayout}} used to display the scale.}
\usage{gtkScaleGetLayout(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkScale}}]  A \verb{\link{GtkScale}}}}
\details{  Since  2.4}
\value{[\verb{\link{PangoLayout}}]  the \verb{\link{PangoLayout}} for this scale, or \code{NULL} 
   if the \verb{"draw-value"} property is \code{FALSE}.
  }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
