\alias{gtkScaleGetValuePos}
\name{gtkScaleGetValuePos}
\title{gtkScaleGetValuePos}
\description{Gets the position in which the current value is displayed.}
\usage{gtkScaleGetValuePos(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkScale}}}}
\value{[\code{\link{GtkPositionType}}] the position in which the current value is displayed}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
