\alias{gtkScaleSetDigits}
\name{gtkScaleSetDigits}
\title{gtkScaleSetDigits}
\description{Sets the number of decimal places that are displayed in the value.
Also causes the value of the adjustment to be rounded off to this
number of digits, so the retrieved value matches the value the user saw.}
\usage{gtkScaleSetDigits(object, digits)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkScale}}}
\item{\verb{digits}}{the number of decimal places to display,
e.g. use 1 to display 1.0, 2 to display 1.00, etc}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
