\alias{gtkScaleSetDrawValue}
\name{gtkScaleSetDrawValue}
\title{gtkScaleSetDrawValue}
\description{Specifies whether the current value is displayed as a string next to the
slider.}
\usage{gtkScaleSetDrawValue(object, draw.value)}
\arguments{
\item{\code{object}}{[\code{\link{GtkScale}}] a \code{\link{GtkScale}}.}
\item{\code{draw.value}}{[logical] a boolean.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
