\alias{gtkScaleSetValuePos}
\name{gtkScaleSetValuePos}
\title{gtkScaleSetValuePos}
\description{Sets the position in which the current value is displayed.}
\usage{gtkScaleSetValuePos(object, pos)}
\arguments{
\item{\code{object}}{[\code{\link{GtkScale}}] a \code{\link{GtkScale}}.}
\item{\code{pos}}{[\code{\link{GtkPositionType}}] the position in which the current value is displayed.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
