\alias{gtkScrolledWindowGetHadjustment}
\name{gtkScrolledWindowGetHadjustment}
\title{gtkScrolledWindowGetHadjustment}
\description{Returns the horizontal scrollbar's adjustment, used to connect the
horizontal scrollbar to the child widget's horizontal scroll
functionality.}
\usage{gtkScrolledWindowGetHadjustment(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkScrolledWindow}}] A \code{\link{GtkScrolledWindow}}.}}
\value{[\code{\link{GtkAdjustment}}] The horizontal \code{\link{GtkAdjustment}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
