\alias{gtkScrolledWindowGetPlacement}
\name{gtkScrolledWindowGetPlacement}
\title{gtkScrolledWindowGetPlacement}
\description{Gets the placement of the scrollbars for the scrolled window. See 
\code{\link{gtkScrolledWindowSetPlacement}}.}
\usage{gtkScrolledWindowGetPlacement(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkScrolledWindow}}]  a \code{\link{GtkScrolledWindow}}}}
\value{[\code{\link{GtkCornerType}}]  the current placement value.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
