\alias{gtkScrolledWindowGetShadowType}
\name{gtkScrolledWindowGetShadowType}
\title{gtkScrolledWindowGetShadowType}
\description{Gets the shadow type of the scrolled window. See 
\code{\link{gtkScrolledWindowSetShadowType}}.}
\usage{gtkScrolledWindowGetShadowType(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkScrolledWindow}}}}
\value{[\code{\link{GtkShadowType}}] the current shadow type}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
