\alias{gtkScrolledWindowSetHadjustment}
\name{gtkScrolledWindowSetHadjustment}
\title{gtkScrolledWindowSetHadjustment}
\description{Sets the \code{\link{GtkAdjustment}} for the horizontal scrollbar.}
\usage{gtkScrolledWindowSetHadjustment(object, hadjustment)}
\arguments{
\item{\code{object}}{[\code{\link{GtkScrolledWindow}}] A \code{\link{GtkScrolledWindow}}.}
\item{\code{hadjustment}}{[\code{\link{GtkAdjustment}}] Horizontal scroll adjustment.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
