\alias{gtkScrolledWindowSetShadowType}
\name{gtkScrolledWindowSetShadowType}
\title{gtkScrolledWindowSetShadowType}
\description{Changes the type of shadow drawn around the contents of
\code{scrolled.window}.}
\usage{gtkScrolledWindowSetShadowType(object, type)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkScrolledWindow}}}
\item{\verb{type}}{kind of shadow to draw around scrolled window contents}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
