\alias{gtkScrolledWindowSetVadjustment}
\name{gtkScrolledWindowSetVadjustment}
\title{gtkScrolledWindowSetVadjustment}
\description{Sets the \code{\link{GtkAdjustment}} for the vertical scrollbar.}
\usage{gtkScrolledWindowSetVadjustment(object, hadjustment)}
\arguments{
\item{\code{object}}{[\code{\link{GtkScrolledWindow}}] A \code{\link{GtkScrolledWindow}}.}
\item{\code{hadjustment}}{[\code{\link{GtkAdjustment}}] Vertical scroll adjustment.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
