\alias{gtkSelectionAddTargets}
\name{gtkSelectionAddTargets}
\title{gtkSelectionAddTargets}
\description{Prepends a table of targets to the list of supported targets
for a given widget and selection.}
\usage{gtkSelectionAddTargets(object, selection, targets)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}}
\item{\verb{selection}}{[\verb{\link{GdkAtom}}]  the selection}
\item{\verb{targets}}{[\verb{\link{GtkTargetEntry}}]  a table of targets to add}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
