\alias{gtkSelectionClearTargets}
\name{gtkSelectionClearTargets}
\title{gtkSelectionClearTargets}
\description{Remove all targets registered for the given selection for the
widget.}
\usage{gtkSelectionClearTargets(object, selection)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkWidget}}]     a \verb{\link{GtkWidget}}}
\item{\verb{selection}}{[\verb{\link{GdkAtom}}]  an atom representing a selection}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
