\alias{gtkSelectionDataCopy}
\name{gtkSelectionDataCopy}
\title{gtkSelectionDataCopy}
\description{Makes a copy of a \verb{\link{GtkSelectionData}} structure and its data.}
\usage{gtkSelectionDataCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkSelectionData}}]  a pointer to a \verb{\link{GtkSelectionData}} structure.}}
\value{[\verb{\link{GtkSelectionData}}]  a pointer to a copy of \code{data}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
