\alias{gtkSelectionDataGetData}
\name{gtkSelectionDataGetData}
\title{gtkSelectionDataGetData}
\description{Retrieves the raw data of the selection.}
\usage{gtkSelectionDataGetData(object)}
\arguments{\item{\verb{object}}{a pointer to a \code{\link{GtkSelectionData}} structure.}}
\details{Since 2.14}
\value{[raw] the raw data of the selection.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
