\alias{gtkSelectionDataGetDataType}
\name{gtkSelectionDataGetDataType}
\title{gtkSelectionDataGetDataType}
\description{Retrieves the data type of the selection.}
\usage{gtkSelectionDataGetDataType(object)}
\arguments{\item{\verb{object}}{a pointer to a \code{\link{GtkSelectionData}} structure.}}
\details{Since 2.14}
\value{[\code{\link{GdkAtom}}] the data type of the selection.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
