\alias{gtkSelectionDataGetDisplay}
\name{gtkSelectionDataGetDisplay}
\title{gtkSelectionDataGetDisplay}
\description{Retrieves the display of the selection.}
\usage{gtkSelectionDataGetDisplay(object)}
\arguments{\item{\verb{object}}{a pointer to a \code{\link{GtkSelectionData}} structure.}}
\details{Since 2.14}
\value{[\code{\link{GdkDisplay}}] the display of the selection.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
