\alias{gtkSelectionDataGetFormat}
\name{gtkSelectionDataGetFormat}
\title{gtkSelectionDataGetFormat}
\description{Retrieves the format of the selection.}
\usage{gtkSelectionDataGetFormat(object)}
\arguments{\item{\verb{object}}{a pointer to a \code{\link{GtkSelectionData}} structure.}}
\details{Since 2.14}
\value{[integer] the format of the selection.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
