\alias{gtkSelectionDataGetLength}
\name{gtkSelectionDataGetLength}
\title{gtkSelectionDataGetLength}
\description{Retrieves the length of the raw data of the selection.}
\usage{gtkSelectionDataGetLength(object)}
\arguments{\item{\verb{object}}{a pointer to a \code{\link{GtkSelectionData}} structure.}}
\details{Since 2.14}
\value{[integer] the length of the data of the selection.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
