\alias{gtkSelectionDataGetSelection}
\name{gtkSelectionDataGetSelection}
\title{gtkSelectionDataGetSelection}
\description{Retrieves the selection \code{\link{GdkAtom}} of the selection data.}
\usage{gtkSelectionDataGetSelection(object)}
\arguments{\item{\verb{object}}{a pointer to a \code{\link{GtkSelectionData}} structure.}}
\details{Since 2.16}
\value{[\code{\link{GdkAtom}}] the selection \code{\link{GdkAtom}} of the selection data.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
