\alias{gtkSelectionDataGetTarget}
\name{gtkSelectionDataGetTarget}
\title{gtkSelectionDataGetTarget}
\description{Retrieves the target of the selection.}
\usage{gtkSelectionDataGetTarget(object)}
\arguments{\item{\verb{object}}{a pointer to a \code{\link{GtkSelectionData}} structure.}}
\details{Since 2.14}
\value{[\code{\link{GdkAtom}}] the target of the selection.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
