\alias{gtkSelectionRemoveAll}
\name{gtkSelectionRemoveAll}
\title{gtkSelectionRemoveAll}
\description{Removes all handlers and unsets ownership of all 
selections for a widget. Called when widget is being
destroyed. This function will not generally be
called by applications.}
\usage{gtkSelectionRemoveAll(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
