\alias{GtkSeparator}
\name{GtkSeparator}
\title{GtkSeparator}
\description{Base class for    and }
\details{The \code{\link{GtkSeparator}} widget is an abstract class, used only for deriving the
subclasses \code{\link{GtkHSeparator}} and \code{\link{GtkVSeparator}}.}
\section{Hierarchy}{\preformatted{  \link{GObject}
   +----\link{GtkObject}
         +----\link{GtkWidget}
               +----GtkSeparator
                     +----\link{GtkHSeparator}
                     +----\link{GtkVSeparator}}}
\section{Structures}{\describe{\item{\code{GtkSeparator}}{
The \code{\link{GtkSeparator}} struct contains private data only.

}}}
\section{Interfaces}{GtkSeparator implements
 AtkImplementorIface.}
\references{\url{http://developer.gnome.org/doc/API/2.0/gtk/GtkSeparator.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
