\alias{gtkSeparatorMenuItemNew}
\name{gtkSeparatorMenuItemNew}
\title{gtkSeparatorMenuItemNew}
\description{Creates a new \code{\link{GtkSeparatorMenuItem}}.}
\usage{gtkSeparatorMenuItemNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkSeparatorMenuItem}}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
