\alias{gtkSeparatorToolItemGetDraw}
\name{gtkSeparatorToolItemGetDraw}
\title{gtkSeparatorToolItemGetDraw}
\description{Returns whether \code{separator.tool.item} is drawn as a
line, or just blank. See \code{\link{gtkSeparatorToolItemSetDraw}}.}
\usage{gtkSeparatorToolItemGetDraw(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkSeparatorToolItem}}]  a \verb{\link{GtkSeparatorToolItem}} }}
\details{  Since  2.4}
\value{[logical] \verb{TRUE} if \code{separator.tool.item} is drawn as a line, or just blank.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
