\alias{gtkSeparatorToolItemNew}
\name{gtkSeparatorToolItemNew}
\title{gtkSeparatorToolItemNew}
\description{Create a new \verb{\link{GtkSeparatorToolItem}}}
\usage{gtkSeparatorToolItemNew(show = TRUE)}
\details{  Since  2.4}
\value{[\verb{\link{GtkToolItem}}]  the new \verb{\link{GtkSeparatorToolItem}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
