\alias{gtkSeparatorToolItemSetDraw}
\name{gtkSeparatorToolItemSetDraw}
\title{gtkSeparatorToolItemSetDraw}
\description{When \code{separator.tool.items} is drawn as a vertical line, or just blank.
Setting this \code{FALSE} along with \code{\link{gtkToolItemSetExpand}} is useful
to create an item that forces following items to the end of the toolbar.}
\usage{gtkSeparatorToolItemSetDraw(object, draw)}
\arguments{
\item{\code{object}}{[\code{\link{GtkSeparatorToolItem}}]  a \code{\link{GtkSeparatorToolItem}}}
\item{\code{draw}}{[logical]  whether \code{separator.tool.item} is drawn as a vertical line}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
