\alias{gtkSettingsGetDefault}
\name{gtkSettingsGetDefault}
\title{gtkSettingsGetDefault}
\description{Gets the \verb{\link{GtkSettings}} object for the default GDK screen, creating
it if necessary. See \code{\link{gtkSettingsGetForScreen}}.}
\usage{gtkSettingsGetDefault()}
\value{[\verb{\link{GtkSettings}}]  a \verb{\link{GtkSettings}} object. If there is no default
 screen, then returns \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
