\alias{gtkSettingsGetForScreen}
\name{gtkSettingsGetForScreen}
\title{gtkSettingsGetForScreen}
\description{Gets the \verb{\link{GtkSettings}} object for \code{screen}, creating it if necessary.}
\usage{gtkSettingsGetForScreen(screen)}
\arguments{\item{\verb{screen}}{[\verb{\link{GdkScreen}}]  a \verb{\link{GdkScreen}}.}}
\details{  Since  2.2}
\value{[\verb{\link{GtkSettings}}]  a \verb{\link{GtkSettings}} object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
