\alias{gtkSettingsInstallPropertyParser}
\name{gtkSettingsInstallPropertyParser}
\title{gtkSettingsInstallPropertyParser}
\description{\emph{undocumented
}}
\usage{gtkSettingsInstallPropertyParser(pspec, parser)}
\arguments{
\item{\code{pspec}}{[\code{\link{GParamSpec}}] }
\item{\code{parser}}{[GtkRcPropertyParser] }
}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
