\alias{gtkSettingsSetDoubleProperty}
\name{gtkSettingsSetDoubleProperty}
\title{gtkSettingsSetDoubleProperty}
\description{\emph{undocumented
}}
\usage{gtkSettingsSetDoubleProperty(object, name, v.double, origin)}
\arguments{
\item{\code{object}}{[\code{\link{GtkSettings}}] }
\item{\code{name}}{[character] }
\item{\code{v.double}}{[numeric] }
\item{\code{origin}}{[character] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
