\alias{gtkSettingsSetLongProperty}
\name{gtkSettingsSetLongProperty}
\title{gtkSettingsSetLongProperty}
\description{\emph{undocumented
}}
\usage{gtkSettingsSetLongProperty(object, name, v.long, origin)}
\arguments{
\item{\code{object}}{[\code{\link{GtkSettings}}] }
\item{\code{name}}{[character] }
\item{\code{v.long}}{[numeric] }
\item{\code{origin}}{[character] }
}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
