\alias{gtkSettingsSetPropertyValue}
\name{gtkSettingsSetPropertyValue}
\title{gtkSettingsSetPropertyValue}
\usage{gtkSettingsSetPropertyValue(object, name, svalue)}
\arguments{
\item{\code{object}}{[\code{\link{GtkSettings}}] }
\item{\code{name}}{[character] }
\item{\code{svalue}}{[\code{\link{GtkSettingsValue}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
