\alias{gtkSettingsSetStringProperty}
\name{gtkSettingsSetStringProperty}
\title{gtkSettingsSetStringProperty}
\description{\emph{undocumented
}}
\usage{gtkSettingsSetStringProperty(object, name, v.string, origin)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkSettings}}] }
\item{\verb{name}}{[character] }
\item{\verb{v.string}}{[character] }
\item{\verb{origin}}{[character] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
