\alias{gtkShowAboutDialog}
\name{gtkShowAboutDialog}
\title{gtkShowAboutDialog}
\description{This is a convenience function for showing an application's about box.
The constructed dialog is associated with the parent window and 
reused for future invocations of this function.}
\usage{gtkShowAboutDialog(parent, ...)}
\arguments{\item{\code{...}}{[character]  the name of the first property }}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
