\alias{gtkSizeGroupGetMode}
\name{gtkSizeGroupGetMode}
\title{gtkSizeGroupGetMode}
\description{Gets the current mode of the size group. See \code{\link{gtkSizeGroupSetMode}}.}
\usage{gtkSizeGroupGetMode(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkSizeGroup}}}}
\value{[\code{\link{GtkSizeGroupMode}}] the current mode of the size group.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
