\alias{gtkSizeGroupRemoveWidget}
\name{gtkSizeGroupRemoveWidget}
\title{gtkSizeGroupRemoveWidget}
\description{Removes a widget from a \verb{\link{GtkSizeGroup}}.}
\usage{gtkSizeGroupRemoveWidget(object, widget)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkSizeGroup}}]  a \verb{GtkSizeGrup}}
\item{\verb{widget}}{[\verb{\link{GtkWidget}}]  the \verb{\link{GtkWidget}} to remove}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
