\alias{gtkSocketGetPlugWindow}
\name{gtkSocketGetPlugWindow}
\title{gtkSocketGetPlugWindow}
\description{Retrieves the window of the plug. Use this to check if the plug has
been created inside of the socket.}
\usage{gtkSocketGetPlugWindow(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkSocket}}.}}
\details{Since 2.14}
\value{[\code{\link{GdkWindow}}] the window of the plug if available, or \code{NULL}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
