\alias{gtkSpinButtonGetAdjustment}
\name{gtkSpinButtonGetAdjustment}
\title{gtkSpinButtonGetAdjustment}
\description{Get the adjustment associated with a \verb{\link{GtkSpinButton}}}
\usage{gtkSpinButtonGetAdjustment(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkSpinButton}}]  a \verb{\link{GtkSpinButton}}}}
\value{[\verb{\link{GtkAdjustment}}]  the \verb{\link{GtkAdjustment}} of \code{spin.button}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
