\alias{gtkSpinButtonGetDigits}
\name{gtkSpinButtonGetDigits}
\title{gtkSpinButtonGetDigits}
\description{Fetches the precision of \code{spin.button}. See \code{\link{gtkSpinButtonSetDigits}}.}
\usage{gtkSpinButtonGetDigits(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkSpinButton}}}}
\value{[numeric] the current precision}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
