\alias{gtkSpinButtonGetNumeric}
\name{gtkSpinButtonGetNumeric}
\title{gtkSpinButtonGetNumeric}
\description{Returns whether non-numeric text can be typed into the spin button.
See \code{\link{gtkSpinButtonSetNumeric}}.}
\usage{gtkSpinButtonGetNumeric(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkSpinButton}}}}
\value{[logical] \code{TRUE} if only numeric text can be entered}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
