\alias{gtkSpinButtonGetSnapToTicks}
\name{gtkSpinButtonGetSnapToTicks}
\title{gtkSpinButtonGetSnapToTicks}
\description{Returns whether the values are corrected to the nearest step. See
\code{\link{gtkSpinButtonSetSnapToTicks}}.}
\usage{gtkSpinButtonGetSnapToTicks(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkSpinButton}}}}
\value{[logical] \code{TRUE} if values are snapped to the nearest step.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
