\alias{gtkSpinButtonGetUpdatePolicy}
\name{gtkSpinButtonGetUpdatePolicy}
\title{gtkSpinButtonGetUpdatePolicy}
\description{Gets the update behavior of a spin button. See
\code{\link{gtkSpinButtonSetUpdatePolicy}}.}
\usage{gtkSpinButtonGetUpdatePolicy(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkSpinButton}}}}
\value{[\code{\link{GtkSpinButtonUpdatePolicy}}] the current update policy}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
