\alias{gtkSpinButtonGetValue}
\name{gtkSpinButtonGetValue}
\title{gtkSpinButtonGetValue}
\description{Get the value in the \code{spin.button}.}
\usage{gtkSpinButtonGetValue(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkSpinButton}}}}
\value{[numeric] the value of \code{spin.button}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
