\alias{gtkSpinButtonGetValueAsInt}
\name{gtkSpinButtonGetValueAsInt}
\title{gtkSpinButtonGetValueAsInt}
\description{Get the value \code{spin.button} represented as an integer.}
\usage{gtkSpinButtonGetValueAsInt(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkSpinButton}}]  a \code{\link{GtkSpinButton}}}}
\value{[integer]  the value of \code{spin.button}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
