\alias{gtkSpinButtonGetWrap}
\name{gtkSpinButtonGetWrap}
\title{gtkSpinButtonGetWrap}
\description{Returns whether the spin button's value wraps around to the
opposite limit when the upper or lower limit of the range is
exceeded. See \code{\link{gtkSpinButtonSetWrap}}.}
\usage{gtkSpinButtonGetWrap(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkSpinButton}}}}
\value{[logical] \code{TRUE} if the spin button wraps around}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
