\alias{gtkSpinButtonNew}
\name{gtkSpinButtonNew}
\title{gtkSpinButtonNew}
\description{Creates a new \verb{\link{GtkSpinButton}}.}
\usage{gtkSpinButtonNew(adjustment = NULL, climb.rate = NULL, digits = NULL, show = TRUE)}
\arguments{
\item{\verb{adjustment}}{[\verb{\link{GtkAdjustment}}] the \verb{\link{GtkAdjustment}} object that this spin button should use.}
\item{\verb{climb.rate}}{[numeric] specifies how much the spin button changes when an arrow is clicked on.}
\item{\verb{digits}}{[numeric] the number of decimal places to display.}
}
\value{[\verb{\link{GtkWidget}}] The new spin button as a \verb{\link{GtkWidget}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
