\alias{gtkSpinButtonSetAdjustment}
\name{gtkSpinButtonSetAdjustment}
\title{gtkSpinButtonSetAdjustment}
\description{Replaces the \verb{\link{GtkAdjustment}} associated with \code{spin.button}.}
\usage{gtkSpinButtonSetAdjustment(object, adjustment)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkSpinButton}}]  a \verb{\link{GtkSpinButton}}}
\item{\verb{adjustment}}{[\verb{\link{GtkAdjustment}}]  a \verb{\link{GtkAdjustment}} to replace the existing adjustment}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
