\alias{gtkSpinButtonSetDigits}
\name{gtkSpinButtonSetDigits}
\title{gtkSpinButtonSetDigits}
\description{Set the precision to be displayed by \code{spin.button}. Up to 20 digit precision
is allowed.}
\usage{gtkSpinButtonSetDigits(object, digits)}
\arguments{
\item{\code{object}}{[\code{\link{GtkSpinButton}}]  a \code{\link{GtkSpinButton}}}
\item{\code{digits}}{[numeric]  the number of digits after the decimal point to be displayed for the spin button's value}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
