\alias{gtkSpinButtonSetIncrements}
\name{gtkSpinButtonSetIncrements}
\title{gtkSpinButtonSetIncrements}
\description{Sets the step and page increments for spin\_button.  This affects how 
quickly the value changes when the spin button's arrows are activated.}
\usage{gtkSpinButtonSetIncrements(object, step, page)}
\arguments{
\item{\code{object}}{[\code{\link{GtkSpinButton}}]  a \code{\link{GtkSpinButton}}}
\item{\code{step}}{[numeric]  increment applied for a button 1 press.}
\item{\code{page}}{[numeric]  increment applied for a button 2 press.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
