\alias{gtkSpinButtonSetNumeric}
\name{gtkSpinButtonSetNumeric}
\title{gtkSpinButtonSetNumeric}
\description{Sets the flag that determines if non-numeric text can be typed into
the spin button.}
\usage{gtkSpinButtonSetNumeric(object, numeric)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkSpinButton}}]  a \verb{\link{GtkSpinButton}} }
\item{\verb{numeric}}{[logical]  flag indicating if only numeric entry is allowed. }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
