\alias{gtkSpinButtonSetSnapToTicks}
\name{gtkSpinButtonSetSnapToTicks}
\title{gtkSpinButtonSetSnapToTicks}
\description{Sets the policy as to whether values are corrected to the nearest step 
increment when a spin button is activated after providing an invalid value.}
\usage{gtkSpinButtonSetSnapToTicks(object, snap.to.ticks)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkSpinButton}}]  a \verb{\link{GtkSpinButton}} }
\item{\verb{snap.to.ticks}}{[logical]  a flag indicating if invalid values should be corrected.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
