\alias{gtkSpinButtonSetValue}
\name{gtkSpinButtonSetValue}
\title{gtkSpinButtonSetValue}
\description{Set the value of \code{spin.button}.}
\usage{gtkSpinButtonSetValue(object, value)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkSpinButton}}]  a \verb{\link{GtkSpinButton}}}
\item{\verb{value}}{[numeric]  the new value}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
