\alias{gtkSpinButtonSetWrap}
\name{gtkSpinButtonSetWrap}
\title{gtkSpinButtonSetWrap}
\description{Sets the flag that determines if a spin button value wraps around to the
opposite limit when the upper or lower limit of the range is exceeded.}
\usage{gtkSpinButtonSetWrap(object, wrap)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkSpinButton}}]  a \verb{\link{GtkSpinButton}} }
\item{\verb{wrap}}{[logical]  a flag indicating if wrapping behavior is performed.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
