\alias{gtkSpinButtonSpin}
\name{gtkSpinButtonSpin}
\title{gtkSpinButtonSpin}
\description{Increment or decrement a spin button's value in a specified direction
by a specified amount.}
\usage{gtkSpinButtonSpin(object, direction, increment)}
\arguments{
\item{\code{object}}{[\code{\link{GtkSpinButton}}]  a \code{\link{GtkSpinButton}} }
\item{\code{direction}}{[\code{\link{GtkSpinType}}]  a \code{\link{GtkSpinType}} indicating the direction to spin.}
\item{\code{increment}}{[numeric]  step increment to apply in the specified direction.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
