\alias{gtkSpinButtonUpdate}
\name{gtkSpinButtonUpdate}
\title{gtkSpinButtonUpdate}
\description{Manually force an update of the spin button.}
\usage{gtkSpinButtonUpdate(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkSpinButton}}]  a \code{\link{GtkSpinButton}} }}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
