\alias{gtkSpinnerNew}
\name{gtkSpinnerNew}
\title{gtkSpinnerNew}
\description{Returns a new spinner widget. Not yet started.}
\usage{gtkSpinnerNew(show = TRUE)}
\details{Since 2.20}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkSpinner}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
