\alias{gtkSpinnerStart}
\name{gtkSpinnerStart}
\title{gtkSpinnerStart}
\description{Starts the animation of the spinner.}
\usage{gtkSpinnerStart(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkSpinner}}}}
\details{Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
