\alias{gtkSpinnerStop}
\name{gtkSpinnerStop}
\title{gtkSpinnerStop}
\description{Stops the animation of the spinner.}
\usage{gtkSpinnerStop(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkSpinner}}}}
\details{Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
