\alias{gtkStatusbarGetContextId}
\name{gtkStatusbarGetContextId}
\title{gtkStatusbarGetContextId}
\description{Returns a new context identifier, given a description 
of the actual context. Note that the description is 
\emph{not} shown in the UI.}
\usage{gtkStatusbarGetContextId(object, context.description)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkStatusbar}}}
\item{\verb{context.description}}{textual description of what context
the new message is being used in}
}
\value{[numeric] an integer id}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
