\alias{gtkStatusbarGetMessageArea}
\name{gtkStatusbarGetMessageArea}
\title{gtkStatusbarGetMessageArea}
\description{Retrieves the box containing the label widget.}
\usage{gtkStatusbarGetMessageArea(object)}
\arguments{\item{\verb{object}}{a \verb{GtkStatusBar}}}
\details{Since 2.20}
\value{[\code{\link{GtkWidget}}] a \code{\link{GtkBox}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
