\alias{gtkStatusbarNew}
\name{gtkStatusbarNew}
\title{gtkStatusbarNew}
\description{Creates a new \code{\link{GtkStatusbar}} ready for messages.}
\usage{gtkStatusbarNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}]  the new \code{\link{GtkStatusbar}}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
