\alias{gtkStatusIconGetBlinking}
\name{gtkStatusIconGetBlinking}
\title{gtkStatusIconGetBlinking}
\description{Returns whether the icon is blinking, see 
\code{\link{gtkStatusIconSetBlinking}}.}
\usage{gtkStatusIconGetBlinking(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkStatusIcon}}]  a \verb{\link{GtkStatusIcon}}}}
\details{  Since  2.10}
\value{[logical] \code{TRUE} if the icon is blinking}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
