\alias{gtkStatusIconGetHasTooltip}
\name{gtkStatusIconGetHasTooltip}
\title{gtkStatusIconGetHasTooltip}
\description{Returns the current value of the has-tooltip property.
See \verb{"has-tooltip"} for more information.}
\usage{gtkStatusIconGetHasTooltip(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkStatusIcon}}}}
\details{Since 2.16}
\value{[logical] current value of has-tooltip on \code{status.icon}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
