\alias{gtkStatusIconGetScreen}
\name{gtkStatusIconGetScreen}
\title{gtkStatusIconGetScreen}
\description{Returns the \verb{\link{GdkScreen}} associated with \code{status.icon}.}
\usage{gtkStatusIconGetScreen(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkStatusIcon}}]  a \verb{\link{GtkStatusIcon}}}}
\details{  Since  2.12}
\value{[\verb{\link{GdkScreen}}]  a \verb{\link{GdkScreen}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
