\alias{gtkStatusIconGetStorageType}
\name{gtkStatusIconGetStorageType}
\title{gtkStatusIconGetStorageType}
\description{Gets the type of representation being used by the \verb{\link{GtkStatusIcon}}
to store image data. If the \verb{\link{GtkStatusIcon}} has no image data,
the return value will be \code{GTK_IMAGE_EMPTY}.}
\usage{gtkStatusIconGetStorageType(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkStatusIcon}}]  a \verb{\link{GtkStatusIcon}}}}
\details{  Since  2.10}
\value{[\code{\link{GtkImageType}}]  the image representation being used}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
