\alias{gtkStatusIconGetTitle}
\name{gtkStatusIconGetTitle}
\title{gtkStatusIconGetTitle}
\description{Gets the title of this tray icon. See \code{\link{gtkStatusIconSetTitle}}.}
\usage{gtkStatusIconGetTitle(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkStatusIcon}}}}
\details{Since 2.18}
\value{[character] the title of the status icon}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
