\alias{gtkStatusIconGetTooltipText}
\name{gtkStatusIconGetTooltipText}
\title{gtkStatusIconGetTooltipText}
\description{Gets the contents of the tooltip for \code{status.icon}.}
\usage{gtkStatusIconGetTooltipText(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkStatusIcon}}}}
\details{Since 2.16}
\value{[character] the tooltip text, or \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
