\alias{gtkStatusIconGetVisible}
\name{gtkStatusIconGetVisible}
\title{gtkStatusIconGetVisible}
\description{Returns whether the status icon is visible or not. 
Note that being visible does not guarantee that 
the user can actually see the icon, see also 
\code{\link{gtkStatusIconIsEmbedded}}.}
\usage{gtkStatusIconGetVisible(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkStatusIcon}}}}
\details{Since 2.10}
\value{[logical] \code{TRUE} if the status icon is visible}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
