\alias{gtkStatusIconIsEmbedded}
\name{gtkStatusIconIsEmbedded}
\title{gtkStatusIconIsEmbedded}
\description{Returns whether the status icon is embedded in a notification
area.}
\usage{gtkStatusIconIsEmbedded(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkStatusIcon}}}}
\details{Since 2.10}
\value{[logical] \code{TRUE} if the status icon is embedded in
a notification area.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
