\alias{gtkStatusIconNew}
\name{gtkStatusIconNew}
\title{gtkStatusIconNew}
\description{Creates an empty status icon object.}
\usage{gtkStatusIconNew()}
\details{  Since  2.10}
\value{[\verb{\link{GtkStatusIcon}}]  a new \verb{\link{GtkStatusIcon}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
