\alias{gtkStatusIconNewFromFile}
\name{gtkStatusIconNewFromFile}
\title{gtkStatusIconNewFromFile}
\description{Creates a status icon displaying the file \code{filename}. }
\usage{gtkStatusIconNewFromFile(filename)}
\arguments{\item{\verb{filename}}{[character]  a filename}}
\details{The image will be scaled down to fit in the available 
space in the notification area, if necessary.  

  Since  2.10}
\value{[\verb{\link{GtkStatusIcon}}]  a new \verb{\link{GtkStatusIcon}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
