\alias{gtkStatusIconNewFromGicon}
\name{gtkStatusIconNewFromGicon}
\title{gtkStatusIconNewFromGicon}
\description{Creates a status icon displaying a \code{\link{GIcon}}. If the icon is a
themed icon, it will be updated when the theme changes.}
\usage{gtkStatusIconNewFromGicon(icon)}
\arguments{\item{\verb{icon}}{a \code{\link{GIcon}}}}
\details{Since 2.14}
\value{[\code{\link{GtkStatusIcon}}] a new \code{\link{GtkStatusIcon}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
