\alias{gtkStatusIconNewFromIconName}
\name{gtkStatusIconNewFromIconName}
\title{gtkStatusIconNewFromIconName}
\description{Creates a status icon displaying an icon from the current icon theme.
If the current icon theme is changed, the icon will be updated 
appropriately.}
\usage{gtkStatusIconNewFromIconName(icon.name)}
\arguments{\item{\verb{icon.name}}{[character]  an icon name}}
\details{  Since  2.10}
\value{[\verb{\link{GtkStatusIcon}}]  a new \verb{\link{GtkStatusIcon}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
