\alias{gtkStatusIconSetBlinking}
\name{gtkStatusIconSetBlinking}
\title{gtkStatusIconSetBlinking}
\description{Makes the status icon start or stop blinking. 
Note that blinking user interface elements may be problematic
for some users, and thus may be turned off, in which case
this setting has no effect.}
\usage{gtkStatusIconSetBlinking(object, blinking)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkStatusIcon}}}
\item{\verb{blinking}}{\code{TRUE} to turn blinking on, \code{FALSE} to turn it off}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
