\alias{gtkStatusIconSetFromGicon}
\name{gtkStatusIconSetFromGicon}
\title{gtkStatusIconSetFromGicon}
\description{Makes \code{status.icon} display the \code{\link{GIcon}}.
See \code{\link{gtkStatusIconNewFromGicon}} for details.}
\usage{gtkStatusIconSetFromGicon(object, icon)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkStatusIcon}}}
\item{\verb{icon}}{a GIcon}
}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
