\alias{gtkStatusIconSetFromIconName}
\name{gtkStatusIconSetFromIconName}
\title{gtkStatusIconSetFromIconName}
\description{Makes \code{status.icon} display the icon named \code{icon.name} from the 
current icon theme.
See \code{\link{gtkStatusIconNewFromIconName}} for details.}
\usage{gtkStatusIconSetFromIconName(object, icon.name)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkStatusIcon}}}
\item{\verb{icon.name}}{an icon name}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
