\alias{gtkStatusIconSetFromStock}
\name{gtkStatusIconSetFromStock}
\title{gtkStatusIconSetFromStock}
\description{Makes \code{status.icon} display the stock icon with the id \code{stock.id}.
See \code{\link{gtkStatusIconNewFromStock}} for details.}
\usage{gtkStatusIconSetFromStock(object, stock.id)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkStatusIcon}}]  a \verb{\link{GtkStatusIcon}}}
\item{\verb{stock.id}}{[character]  a stock icon id}
}
\details{  Since  2.10 }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
