\alias{gtkStatusIconSetHasTooltip}
\name{gtkStatusIconSetHasTooltip}
\title{gtkStatusIconSetHasTooltip}
\description{Sets the has-tooltip property on \code{status.icon} to \code{has.tooltip}.
See \verb{"has-tooltip"} for more information.}
\usage{gtkStatusIconSetHasTooltip(object, has.tooltip)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkStatusIcon}}}
\item{\verb{has.tooltip}}{whether or not \code{status.icon} has a tooltip}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
