\alias{gtkStatusIconSetName}
\name{gtkStatusIconSetName}
\title{gtkStatusIconSetName}
\description{Sets the name of this tray icon.
This should be a string identifying this icon. It is may be
used for sorting the icons in the tray and will not be shown to
the user.}
\usage{gtkStatusIconSetName(object, name)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkStatusIcon}}}
\item{\verb{name}}{the name}
}
\details{Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
