\alias{gtkStatusIconSetTitle}
\name{gtkStatusIconSetTitle}
\title{gtkStatusIconSetTitle}
\description{Sets the title of this tray icon.
This should be a short, human-readable, localized string 
describing the tray icon. It may be used by tools like screen
readers to render the tray icon.}
\usage{gtkStatusIconSetTitle(object, title)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkStatusIcon}}}
\item{\verb{title}}{the title}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
