\alias{gtkStatusIconSetTooltip}
\name{gtkStatusIconSetTooltip}
\title{gtkStatusIconSetTooltip}
\description{Sets the tooltip of the status icon.}
\usage{gtkStatusIconSetTooltip(object, tooltip.text)}
\arguments{
\item{\code{object}}{[\code{\link{GtkStatusIcon}}]  a \code{\link{GtkStatusIcon}}}
\item{\code{tooltip.text}}{[character]  the tooltip text, or \code{NULL}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
