\alias{gtkStatusIconSetVisible}
\name{gtkStatusIconSetVisible}
\title{gtkStatusIconSetVisible}
\description{Shows or hides a status icon.}
\usage{gtkStatusIconSetVisible(object, visible)}
\arguments{
\item{\code{object}}{[\code{\link{GtkStatusIcon}}]  a \code{\link{GtkStatusIcon}}}
\item{\code{visible}}{[logical] \code{TRUE} to show the status icon, \code{FALSE} to hide it}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
