\alias{gtkStockAddStatic}
\name{gtkStockAddStatic}
\title{gtkStockAddStatic}
\description{Same as \code{\link{gtkStockAdd}}, but doesn't copy \code{items}, so
\code{items} must persist until application exit.}
\usage{gtkStockAddStatic(items)}
\arguments{\item{\code{items}}{[\code{\link{GtkStockItem}}]  a \code{\link{GtkStockItem}} or list of \code{\link{GtkStockItem}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
