\alias{gtkStockItemCopy}
\name{gtkStockItemCopy}
\title{gtkStockItemCopy}
\description{Copies a stock item, mostly useful for language bindings and not in applications.}
\usage{gtkStockItemCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkStockItem}}]  a \verb{\link{GtkStockItem}}}}
\value{[\verb{\link{GtkStockItem}}]  a new \verb{\link{GtkStockItem}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
