\alias{gtkStockListIds}
\name{gtkStockListIds}
\title{gtkStockListIds}
\description{Retrieves a list of all known stock IDs added to a \verb{\link{GtkIconFactory}}
or registered with \code{\link{gtkStockAdd}}.}
\usage{gtkStockListIds()}
\value{[list]  a list of known stock IDs}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
