\alias{gtkStyleApplyDefaultBackground}
\name{gtkStyleApplyDefaultBackground}
\title{gtkStyleApplyDefaultBackground}
\usage{gtkStyleApplyDefaultBackground(object, window, set.bg, state.type, area = NULL, x, y, width, height)}
\arguments{
\item{\code{object}}{[\code{\link{GtkStyle}}] }
\item{\code{window}}{[\code{\link{GdkWindow}}] }
\item{\code{set.bg}}{[logical] }
\item{\code{state.type}}{[\code{\link{GtkStateType}}] }
\item{\code{area}}{[\code{\link{GdkRectangle}}] }
\item{\code{x}}{[integer] }
\item{\code{y}}{[integer] }
\item{\code{width}}{[integer] }
\item{\code{height}}{[integer] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
