\alias{gtkStyleCopy}
\name{gtkStyleCopy}
\title{gtkStyleCopy}
\description{Creates a copy of the passed in \code{\link{GtkStyle}} object.}
\usage{gtkStyleCopy(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkStyle}}}}
\value{[\code{\link{GtkStyle}}] a copy of \code{style}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
