\alias{gtkStyleDetach}
\name{gtkStyleDetach}
\title{gtkStyleDetach}
\description{Detaches a style from a window. If the style is not attached
to any windows anymore, it is unrealized. See \code{\link{gtkStyleAttach}}.}
\usage{gtkStyleDetach(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkStyle}}]  a \verb{\link{GtkStyle}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
