\alias{gtkStyleGet}
\name{gtkStyleGet}
\title{gtkStyleGet}
\description{Gets the values of a multiple style properties for \code{widget.type}
from \code{style}.}
\usage{gtkStyleGet(object, widget.type, first.property.name, ...)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkStyle}}}
\item{\verb{widget.type}}{the \code{\link{GType}} of a descendant of \code{\link{GtkWidget}}}
\item{\verb{first.property.name}}{the name of the first style property to get}
\item{\verb{...}}{\emph{undocumented
}}
}
\details{Since 2.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
