\alias{gtkStyleSetBackground}
\name{gtkStyleSetBackground}
\title{gtkStyleSetBackground}
\description{Sets the background of \code{window} to the background color or pixmap
specified by \code{style} for the given state.}
\usage{gtkStyleSetBackground(object, window, state.type)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkStyle}}]  a \verb{\link{GtkStyle}}}
\item{\verb{window}}{[\verb{\link{GdkWindow}}]  a \verb{\link{GdkWindow}}}
\item{\verb{state.type}}{[\code{\link{GtkStateType}}]  a state}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
