\alias{gtkTableGetColSpacing}
\name{gtkTableGetColSpacing}
\title{gtkTableGetColSpacing}
\description{Gets the amount of space between column \code{col}, and
column \code{col} + 1. See \code{\link{gtkTableSetColSpacing}}.}
\usage{gtkTableGetColSpacing(object, column)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTable}}]  a \verb{\link{GtkTable}}}
\item{\verb{column}}{[numeric]  a column in the table, 0 indicates the first column}
}
\value{[numeric]  the column spacing}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
