\alias{gtkTableGetDefaultColSpacing}
\name{gtkTableGetDefaultColSpacing}
\title{gtkTableGetDefaultColSpacing}
\description{Gets the default column spacing for the table. This is
the spacing that will be used for newly added columns.
(See \code{\link{gtkTableSetColSpacings}})}
\usage{gtkTableGetDefaultColSpacing(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTable}}}}
\value{[numeric] the default column spacing}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
