\alias{gtkTableGetDefaultRowSpacing}
\name{gtkTableGetDefaultRowSpacing}
\title{gtkTableGetDefaultRowSpacing}
\description{Gets the default row spacing for the table. This is
the spacing that will be used for newly added rows.
(See \code{\link{gtkTableSetRowSpacings}})}
\usage{gtkTableGetDefaultRowSpacing(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTable}}]  a \code{\link{GtkTable}}}}
\value{[numeric] value: the default row spacing}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
