\alias{gtkTableGetRowSpacing}
\name{gtkTableGetRowSpacing}
\title{gtkTableGetRowSpacing}
\description{Gets the amount of space between row \code{row}, and
row \code{row} + 1. See \code{\link{gtkTableSetRowSpacing}}.}
\usage{gtkTableGetRowSpacing(object, row)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTable}}]  a \code{\link{GtkTable}}}
\item{\code{row}}{[numeric]  a row in the table, 0 indicates the first row}
}
\value{[numeric]  the row spacing}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
