\alias{gtkTableResize}
\name{gtkTableResize}
\title{gtkTableResize}
\description{If you need to change a table's size \emph{after} it has been created, this function allows you to do so.}
\usage{gtkTableResize(object, rows, columns)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTable}}] The \verb{\link{GtkTable}} you wish to change the size of.}
\item{\verb{rows}}{[numeric] The new number of rows.}
\item{\verb{columns}}{[numeric] The new number of columns.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
