\alias{gtkTableSetColSpacing}
\name{gtkTableSetColSpacing}
\title{gtkTableSetColSpacing}
\description{Alters the amount of space between a given table column and the following 
column.}
\usage{gtkTableSetColSpacing(object, column, spacing)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTable}}] a \verb{\link{GtkTable}}.}
\item{\verb{column}}{[numeric] the column whose spacing should be changed.}
\item{\verb{spacing}}{[numeric] number of pixels that the spacing should take up.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
