\alias{gtkTableSetHomogeneous}
\name{gtkTableSetHomogeneous}
\title{gtkTableSetHomogeneous}
\description{Changes the homogenous property of table cells, ie. whether all cells are an equal size or not.}
\usage{gtkTableSetHomogeneous(object, homogeneous)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkTable}} you wish to set the homogeneous properties of.}
\item{\verb{homogeneous}}{Set to \code{TRUE} to ensure all table cells are the same size. Set
to \code{FALSE} if this is not your desired behaviour.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
