\alias{gtkTableSetRowSpacing}
\name{gtkTableSetRowSpacing}
\title{gtkTableSetRowSpacing}
\description{Changes the space between a given table row and the subsequent row.}
\usage{gtkTableSetRowSpacing(object, row, spacing)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTable}}] a \code{\link{GtkTable}} containing the row whose properties you wish to change.}
\item{\code{row}}{[numeric] row number whose spacing will be changed.}
\item{\code{spacing}}{[numeric] number of pixels that the spacing should take up.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
