\alias{gtkTableSetRowSpacings}
\name{gtkTableSetRowSpacings}
\title{gtkTableSetRowSpacings}
\description{Sets the space between every row in \code{table} equal to \code{spacing}.}
\usage{gtkTableSetRowSpacings(object, spacing)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTable}}] a \code{\link{GtkTable}}.}
\item{\code{spacing}}{[numeric] the number of pixels of space to place between every row in the table.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
