\alias{gtkTargetListAdd}
\name{gtkTargetListAdd}
\title{gtkTargetListAdd}
\description{Appends another target to a \verb{\link{GtkTargetList}}.}
\usage{gtkTargetListAdd(object, target, flags, info)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTargetList}}]   a \verb{\link{GtkTargetList}}}
\item{\verb{target}}{[\verb{\link{GdkAtom}}]  the interned atom representing the target}
\item{\verb{flags}}{[numeric]  the flags for this target}
\item{\verb{info}}{[numeric]  an ID that will be passed back to the application}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
