\alias{gtkTargetListAddTable}
\name{gtkTargetListAddTable}
\title{gtkTargetListAddTable}
\description{Prepends a table of \verb{\link{GtkTargetEntry}} to a target list.}
\usage{gtkTargetListAddTable(object, targets)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTargetList}}]  a \verb{\link{GtkTargetList}}}
\item{\verb{targets}}{[\verb{\link{GtkTargetEntry}}]  the table of \verb{\link{GtkTargetEntry}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
