\alias{gtkTargetListNew}
\name{gtkTargetListNew}
\title{gtkTargetListNew}
\description{Creates a new \code{\link{GtkTargetList}} from a list of \code{\link{GtkTargetEntry}}.}
\usage{gtkTargetListNew(targets)}
\arguments{\item{\code{targets}}{[\code{\link{GtkTargetEntry}}]  Pointer to a list of \code{\link{GtkTargetEntry}}}}
\value{[\code{\link{GtkTargetList}}]  the new \code{\link{GtkTargetList}}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
