\alias{gtkTargetsIncludeText}
\name{gtkTargetsIncludeText}
\title{gtkTargetsIncludeText}
\description{Determines if any of the targets in \code{targets} can be used to
provide text.}
\usage{gtkTargetsIncludeText(targets)}
\arguments{\item{\verb{targets}}{[\verb{\link{GdkAtom}}]  a list of \verb{\link{GdkAtom}}s}}
\details{  Since  2.10}
\value{[logical] \code{TRUE} if \code{targets} include a suitable target for text,
  otherwise \code{FALSE}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
