\alias{gtkTearoffMenuItemNew}
\name{gtkTearoffMenuItemNew}
\title{gtkTearoffMenuItemNew}
\description{Creates a new \verb{\link{GtkTearoffMenuItem}}.}
\usage{gtkTearoffMenuItemNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}] a new \verb{\link{GtkTearoffMenuItem}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
