\alias{gtkTestListAllTypes}
\name{gtkTestListAllTypes}
\title{gtkTestListAllTypes}
\description{Return the type ids that have been registered after
calling \code{\link{gtkTestRegisterAllTypes}}.}
\usage{gtkTestListAllTypes()}
\details{Since 2.14}
\value{
A list containing the following elements:
\item{retval}{[\code{\link{GType}}] array of type ids}
\item{\verb{n.types}}{location to store number of types}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
