\alias{gtkTestRegisterAllTypes}
\name{gtkTestRegisterAllTypes}
\title{gtkTestRegisterAllTypes}
\description{Force registration of all core Gtk+ and Gdk object types.
This allowes to refer to any of those object types via
\code{\link{gTypeFromName}} after calling this function.}
\usage{gtkTestRegisterAllTypes()}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
