\alias{gtkTestTextGet}
\name{gtkTestTextGet}
\title{gtkTestTextGet}
\description{Retrive the text string of \code{widget} if it is a GtkLabel,
GtkEditable (entry and text widgets) or GtkTextView.}
\usage{gtkTestTextGet(widget)}
\arguments{\item{\verb{widget}}{valid widget pointer.}}
\details{Since 2.14}
\value{[character] new C string,}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
