\alias{gtkTestTextSet}
\name{gtkTestTextSet}
\title{gtkTestTextSet}
\description{Set the text string of \code{widget} to \code{string} if it is a GtkLabel,
GtkEditable (entry and text widgets) or GtkTextView.}
\usage{gtkTestTextSet(widget, string)}
\arguments{
\item{\verb{widget}}{valid widget pointer.}
\item{\verb{string}}{a C string}
}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
