\alias{gtkTextAttributesCopy}
\name{gtkTextAttributesCopy}
\title{gtkTextAttributesCopy}
\description{Copies \code{src} and returns a new \verb{\link{GtkTextAttributes}}.}
\usage{gtkTextAttributesCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTextAttributes}}]  a \verb{\link{GtkTextAttributes}} to be copied}}
\value{[\verb{\link{GtkTextAttributes}}]  a copy of \code{src}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
