\alias{gtkTextBufferCopyClipboard}
\name{gtkTextBufferCopyClipboard}
\title{gtkTextBufferCopyClipboard}
\description{Copies the currently-selected text to a clipboard.}
\usage{gtkTextBufferCopyClipboard(object, clipboard)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTextBuffer}}]  a \verb{\link{GtkTextBuffer}} }
\item{\verb{clipboard}}{[\verb{\link{GtkClipboard}}]  the \verb{\link{GtkClipboard}} object to copy to}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
