\alias{gtkTextBufferCutClipboard}
\name{gtkTextBufferCutClipboard}
\title{gtkTextBufferCutClipboard}
\description{Copies the currently-selected text to a clipboard, then deletes
said text if it's editable.}
\usage{gtkTextBufferCutClipboard(object, clipboard, default.editable)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTextBuffer}}]  a \verb{\link{GtkTextBuffer}}}
\item{\verb{clipboard}}{[\verb{\link{GtkClipboard}}]  the \verb{\link{GtkClipboard}} object to cut to}
\item{\verb{default.editable}}{[logical]  default editability of the buffer}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
