\alias{gtkTextBufferEndUserAction}
\name{gtkTextBufferEndUserAction}
\title{gtkTextBufferEndUserAction}
\description{Should be paired with a call to \code{\link{gtkTextBufferBeginUserAction}}.
See that function for a full explanation.}
\usage{gtkTextBufferEndUserAction(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTextBuffer}}]  a \verb{\link{GtkTextBuffer}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
