\alias{gtkTextBufferGetHasSelection}
\name{gtkTextBufferGetHasSelection}
\title{gtkTextBufferGetHasSelection}
\description{Indicates whether the buffer has some text currently selected.}
\usage{gtkTextBufferGetHasSelection(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTextBuffer}}]  a \code{\link{GtkTextBuffer}} }}
\details{  Since  2.10}
\value{[logical]  \code{TRUE} if the there is text selected}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
