\alias{gtkTextBufferGetIterAtMark}
\name{gtkTextBufferGetIterAtMark}
\title{gtkTextBufferGetIterAtMark}
\description{Initializes \code{iter} with the current position of \code{mark}.}
\usage{gtkTextBufferGetIterAtMark(object, mark)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTextBuffer}}]  a \code{\link{GtkTextBuffer}}}
\item{\code{mark}}{[\code{\link{GtkTextMark}}]  a \code{\link{GtkTextMark}} in \code{buffer}}
}
\value{
A list containing the following elements:
\item{\code{iter}}{[\code{\link{GtkTextIter}}]  iterator to initialize}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
