\alias{gtkTextBufferGetLineCount}
\name{gtkTextBufferGetLineCount}
\title{gtkTextBufferGetLineCount}
\description{Obtains the number of lines in the buffer. This value is cached, so
the function is very fast.}
\usage{gtkTextBufferGetLineCount(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextBuffer}}}}
\value{[integer] number of lines in the buffer}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
