\alias{gtkTextBufferMoveMarkByName}
\name{gtkTextBufferMoveMarkByName}
\title{gtkTextBufferMoveMarkByName}
\description{Moves the mark named \code{name} (which must exist) to location \code{where}.
See \code{\link{gtkTextBufferMoveMark}} for details.}
\usage{gtkTextBufferMoveMarkByName(object, name, where)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTextBuffer}}]  a \code{\link{GtkTextBuffer}}}
\item{\code{name}}{[character]  name of a mark}
\item{\code{where}}{[\code{\link{GtkTextIter}}]  new location for mark}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
