\alias{gtkTextBufferRemoveSelectionClipboard}
\name{gtkTextBufferRemoveSelectionClipboard}
\title{gtkTextBufferRemoveSelectionClipboard}
\description{Removes a \verb{\link{GtkClipboard}} added with 
\code{\link{gtkTextBufferAddSelectionClipboard}}.}
\usage{gtkTextBufferRemoveSelectionClipboard(object, clipboard)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTextBuffer}}]  a \verb{\link{GtkTextBuffer}}}
\item{\verb{clipboard}}{[\verb{\link{GtkClipboard}}]  a \verb{\link{GtkClipboard}} added to \code{buffer} by 
            \code{\link{gtkTextBufferAddSelectionClipboard}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
