\alias{gtkTextChildAnchorNew}
\name{gtkTextChildAnchorNew}
\title{gtkTextChildAnchorNew}
\description{Creates a new \verb{\link{GtkTextChildAnchor}}. Usually you would then insert
it into a \verb{\link{GtkTextBuffer}} with \code{\link{gtkTextBufferInsertChildAnchor}}.
To perform the creation and insertion in one step, use the
convenience function \code{\link{gtkTextBufferCreateChildAnchor}}.}
\usage{gtkTextChildAnchorNew()}
\value{[\verb{\link{GtkTextChildAnchor}}]  a new \verb{\link{GtkTextChildAnchor}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
