\alias{gtkTextIterBackwardCursorPosition}
\name{gtkTextIterBackwardCursorPosition}
\title{gtkTextIterBackwardCursorPosition}
\description{Like \code{\link{gtkTextIterForwardCursorPosition}}, but moves backward.}
\usage{gtkTextIterBackwardCursorPosition(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTextIter}}]  a \code{\link{GtkTextIter}}}}
\value{[logical]  \code{TRUE} if we moved}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
