\alias{gtkTextIterBackwardVisibleCursorPosition}
\name{gtkTextIterBackwardVisibleCursorPosition}
\title{gtkTextIterBackwardVisibleCursorPosition}
\description{Moves \code{iter} forward to the previous visible cursor position. See 
\code{\link{gtkTextIterBackwardCursorPosition}} for details.}
\usage{gtkTextIterBackwardVisibleCursorPosition(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextIter}}}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if we moved and the new position is dereferenceable}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
