\alias{gtkTextIterForwardVisibleCursorPosition}
\name{gtkTextIterForwardVisibleCursorPosition}
\title{gtkTextIterForwardVisibleCursorPosition}
\description{Moves \code{iter} forward to the next visible cursor position. See 
\code{\link{gtkTextIterForwardCursorPosition}} for details.}
\usage{gtkTextIterForwardVisibleCursorPosition(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextIter}}}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if we moved and the new position is dereferenceable}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
