\alias{gtkTextIterForwardVisibleWordEnds}
\name{gtkTextIterForwardVisibleWordEnds}
\title{gtkTextIterForwardVisibleWordEnds}
\description{Calls \code{\link{gtkTextIterForwardVisibleWordEnd}} up to \code{count} times.}
\usage{gtkTextIterForwardVisibleWordEnds(object, count)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTextIter}}]  a \code{\link{GtkTextIter}}}
\item{\code{count}}{[integer]  number of times to move}
}
\details{  Since  2.4}
\value{[logical]  \code{TRUE} if \code{iter} moved and is not the end iterator }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
