\alias{gtkTextIterGetBytesInLine}
\name{gtkTextIterGetBytesInLine}
\title{gtkTextIterGetBytesInLine}
\description{Returns the number of bytes in the line containing \code{iter},
including the paragraph delimiters.}
\usage{gtkTextIterGetBytesInLine(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTextIter}}]  an iterator}}
\value{[integer]  number of bytes in the line}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
