\alias{gtkTextIterGetCharsInLine}
\name{gtkTextIterGetCharsInLine}
\title{gtkTextIterGetCharsInLine}
\description{Returns the number of characters in the line containing \code{iter},
including the paragraph delimiters.}
\usage{gtkTextIterGetCharsInLine(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTextIter}}]  an iterator}}
\value{[integer]  number of characters in the line}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
