\alias{gtkTextIterGetLine}
\name{gtkTextIterGetLine}
\title{gtkTextIterGetLine}
\description{Returns the line number containing the iterator. Lines in
a \verb{\link{GtkTextBuffer}} are numbered beginning with 0 for the first
line in the buffer.}
\usage{gtkTextIterGetLine(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTextIter}}]  an iterator}}
\value{[integer]  a line number}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
