\alias{gtkTextIterGetLineOffset}
\name{gtkTextIterGetLineOffset}
\title{gtkTextIterGetLineOffset}
\description{Returns the character offset of the iterator,
counting from the start of a line.
The first character on the line has offset 0.}
\usage{gtkTextIterGetLineOffset(object)}
\arguments{\item{\verb{object}}{an iterator}}
\value{[integer] offset from start of line}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
