\alias{gtkTextIterGetVisibleLineOffset}
\name{gtkTextIterGetVisibleLineOffset}
\title{gtkTextIterGetVisibleLineOffset}
\description{Returns the offset in characters from the start of the
line to the given \code{iter}, not counting characters that
are invisible due to tags with the "invisible" flag
toggled on.}
\usage{gtkTextIterGetVisibleLineOffset(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextIter}}}}
\value{[integer] offset in visible characters from the start of the line}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
