\alias{gtkTextIterIsCursorPosition}
\name{gtkTextIterIsCursorPosition}
\title{gtkTextIterIsCursorPosition}
\description{See \code{\link{gtkTextIterForwardCursorPosition}} or \verb{\link{PangoLogAttr}} or
\code{\link{pangoBreak}} for details on what a cursor position is.}
\usage{gtkTextIterIsCursorPosition(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTextIter}}]  a \verb{\link{GtkTextIter}}}}
\value{[logical] \code{TRUE} if the cursor can be placed at \code{iter}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
