\alias{gtkTextIterIsStart}
\name{gtkTextIterIsStart}
\title{gtkTextIterIsStart}
\description{Returns \code{TRUE} if \code{iter} is the first iterator in the buffer, that is
if \code{iter} has a character offset of 0.}
\usage{gtkTextIterIsStart(object)}
\arguments{\item{\verb{object}}{an iterator}}
\value{[logical] whether \code{iter} is the first in the buffer}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
