\alias{gtkTextIterSetOffset}
\name{gtkTextIterSetOffset}
\title{gtkTextIterSetOffset}
\description{Sets \code{iter} to point to \code{char.offset}. \code{char.offset} counts from the start
of the entire text buffer, starting with 0.}
\usage{gtkTextIterSetOffset(object, char.offset)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextIter}}}
\item{\verb{char.offset}}{a character number}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
