\alias{gtkTextMarkGetBuffer}
\name{gtkTextMarkGetBuffer}
\title{gtkTextMarkGetBuffer}
\description{Gets the buffer this mark is located inside,
or NULL if the mark is deleted.}
\usage{gtkTextMarkGetBuffer(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTextMark}}]  a \code{\link{GtkTextMark}}}}
\value{[\code{\link{GtkTextBuffer}}]  the mark's \code{\link{GtkTextBuffer}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
