\alias{gtkTextMarkGetDeleted}
\name{gtkTextMarkGetDeleted}
\title{gtkTextMarkGetDeleted}
\description{Returns \code{TRUE} if the mark has been removed from its buffer
with \code{\link{gtkTextBufferDeleteMark}}. Marks can't be used
once deleted.}
\usage{gtkTextMarkGetDeleted(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTextMark}}]  a \code{\link{GtkTextMark}}}}
\value{[logical]  whether the mark is deleted}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
