\alias{gtkTextMarkGetLeftGravity}
\name{gtkTextMarkGetLeftGravity}
\title{gtkTextMarkGetLeftGravity}
\description{Determines whether the mark has left gravity.}
\usage{gtkTextMarkGetLeftGravity(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextMark}}}}
\value{[logical] \code{TRUE} if the mark has left gravity, \code{FALSE} otherwise}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
