\alias{gtkTextMarkSetVisible}
\name{gtkTextMarkSetVisible}
\title{gtkTextMarkSetVisible}
\description{Sets the visibility of \code{mark}; the insertion point is normally
visible, i.e. you can see it as a vertical bar. Also, the text
widget uses a visible mark to indicate where a drop will occur when
dragging-and-dropping text. Most other marks are not visible.
Marks are not visible by default.}
\usage{gtkTextMarkSetVisible(object, setting)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextMark}}}
\item{\verb{setting}}{visibility of mark}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
