\alias{gtkTextTagTableGetSize}
\name{gtkTextTagTableGetSize}
\title{gtkTextTagTableGetSize}
\description{Returns the size of the table (number of tags)}
\usage{gtkTextTagTableGetSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextTagTable}}}}
\value{[integer] number of tags in \code{table}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
