\alias{gtkTextTagTableLookup}
\name{gtkTextTagTableLookup}
\title{gtkTextTagTableLookup}
\description{Look up a named tag.}
\usage{gtkTextTagTableLookup(object, name)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTextTagTable}}]  a \verb{\link{GtkTextTagTable}} }
\item{\verb{name}}{[character]  name of a tag}
}
\value{[\verb{\link{GtkTextTag}}]  The tag, or \code{NULL} if none by that name is in the table.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
