\alias{gtkTextTagTableNew}
\name{gtkTextTagTableNew}
\title{gtkTextTagTableNew}
\description{Creates a new \code{\link{GtkTextTagTable}}. The table contains no tags by
default.}
\usage{gtkTextTagTableNew()}
\value{[\code{\link{GtkTextTagTable}}]  a new \code{\link{GtkTextTagTable}}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
