\alias{gtkTextViewAddChildAtAnchor}
\name{gtkTextViewAddChildAtAnchor}
\title{gtkTextViewAddChildAtAnchor}
\description{Adds a child widget in the text buffer, at the given \code{anchor}.}
\usage{gtkTextViewAddChildAtAnchor(object, child, anchor)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTextView}}]  a \verb{\link{GtkTextView}}}
\item{\verb{child}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}}
\item{\verb{anchor}}{[\verb{\link{GtkTextChildAnchor}}]  a \verb{\link{GtkTextChildAnchor}} in the \verb{\link{GtkTextBuffer}} for \code{text.view}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
