\alias{gtkTextViewGetAcceptsTab}
\name{gtkTextViewGetAcceptsTab}
\title{gtkTextViewGetAcceptsTab}
\description{Returns whether pressing the Tab key inserts a tab characters.
\code{\link{gtkTextViewSetAcceptsTab}}.}
\usage{gtkTextViewGetAcceptsTab(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTextView}}}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if pressing the Tab key inserts a tab character,
\code{FALSE} if pressing the Tab key moves the keyboard focus.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
