\alias{gtkTextViewGetBorderWindowSize}
\name{gtkTextViewGetBorderWindowSize}
\title{gtkTextViewGetBorderWindowSize}
\description{Gets the width of the specified border window. See
\code{\link{gtkTextViewSetBorderWindowSize}}.}
\usage{gtkTextViewGetBorderWindowSize(object, type)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTextView}}]  a \verb{\link{GtkTextView}}}
\item{\verb{type}}{[\code{\link{GtkTextWindowType}}]  window to return size from}
}
\value{[integer]  width of window}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
