\alias{gtkTextViewGetCursorVisible}
\name{gtkTextViewGetCursorVisible}
\title{gtkTextViewGetCursorVisible}
\description{Find out whether the cursor is being displayed.}
\usage{gtkTextViewGetCursorVisible(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTextView}}]  a \code{\link{GtkTextView}}}}
\value{[logical]  whether the insertion mark is visible}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
