\alias{gtkTextViewGetEditable}
\name{gtkTextViewGetEditable}
\title{gtkTextViewGetEditable}
\description{Returns the default editability of the \verb{\link{GtkTextView}}. Tags in the
buffer may override this setting for some ranges of text.}
\usage{gtkTextViewGetEditable(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTextView}}]  a \verb{\link{GtkTextView}}}}
\value{[logical]  whether text is editable by default}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
