\alias{gtkTextViewGetIndent}
\name{gtkTextViewGetIndent}
\title{gtkTextViewGetIndent}
\description{Gets the default indentation of paragraphs in \code{text.view}.
Tags in the view's buffer may override the default.
The indentation may be negative.}
\usage{gtkTextViewGetIndent(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextView}}}}
\value{[integer] number of pixels of indentation}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
