\alias{gtkTextViewGetJustification}
\name{gtkTextViewGetJustification}
\title{gtkTextViewGetJustification}
\description{Gets the default justification of paragraphs in \code{text.view}.
Tags in the buffer may override the default.}
\usage{gtkTextViewGetJustification(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextView}}}}
\value{[\code{\link{GtkJustification}}] default justification}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
