\alias{gtkTextViewGetOverwrite}
\name{gtkTextViewGetOverwrite}
\title{gtkTextViewGetOverwrite}
\description{Returns whether the \code{\link{GtkTextView}} is in overwrite mode or not.}
\usage{gtkTextViewGetOverwrite(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTextView}}]  a \code{\link{GtkTextView}}}}
\details{  Since  2.4}
\value{[logical]  whether \code{text.view} is in overwrite mode or not.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
