\alias{gtkTextViewGetPixelsAboveLines}
\name{gtkTextViewGetPixelsAboveLines}
\title{gtkTextViewGetPixelsAboveLines}
\description{Gets the default number of pixels to put above paragraphs.}
\usage{gtkTextViewGetPixelsAboveLines(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextView}}}}
\value{[integer] default number of pixels above paragraphs}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
