\alias{gtkTextViewGetPixelsBelowLines}
\name{gtkTextViewGetPixelsBelowLines}
\title{gtkTextViewGetPixelsBelowLines}
\description{Gets the value set by \code{\link{gtkTextViewSetPixelsBelowLines}}.}
\usage{gtkTextViewGetPixelsBelowLines(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTextView}}]  a \code{\link{GtkTextView}}}}
\value{[integer]  default number of blank pixels below paragraphs}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
