\alias{gtkTextViewGetRightMargin}
\name{gtkTextViewGetRightMargin}
\title{gtkTextViewGetRightMargin}
\description{Gets the default right margin for text in \code{text.view}. Tags
in the buffer may override the default.}
\usage{gtkTextViewGetRightMargin(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextView}}}}
\value{[integer] right margin in pixels}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
