\alias{gtkTextViewGetWrapMode}
\name{gtkTextViewGetWrapMode}
\title{gtkTextViewGetWrapMode}
\description{Gets the line wrapping for the view.}
\usage{gtkTextViewGetWrapMode(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTextView}}]  a \code{\link{GtkTextView}}}}
\value{[\code{\link{GtkWrapMode}}]  the line wrap setting}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
