\alias{gtkTextViewMoveChild}
\name{gtkTextViewMoveChild}
\title{gtkTextViewMoveChild}
\description{Updates the position of a child, as for \code{\link{gtkTextViewAddChildInWindow}}.}
\usage{gtkTextViewMoveChild(object, child, xpos, ypos)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTextView}}]  a \verb{\link{GtkTextView}}}
\item{\verb{child}}{[\verb{\link{GtkWidget}}]  child widget already added to the text view}
\item{\verb{xpos}}{[integer]  new X position in window coordinates}
\item{\verb{ypos}}{[integer]  new Y position in window coordinates}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
