\alias{gtkTextViewSetEditable}
\name{gtkTextViewSetEditable}
\title{gtkTextViewSetEditable}
\description{Sets the default editability of the \verb{\link{GtkTextView}}. You can override
this default setting with tags in the buffer, using the "editable"
attribute of tags.}
\usage{gtkTextViewSetEditable(object, setting)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTextView}}]  a \verb{\link{GtkTextView}}}
\item{\verb{setting}}{[logical]  whether it's editable}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
