\alias{gtkTextViewSetIndent}
\name{gtkTextViewSetIndent}
\title{gtkTextViewSetIndent}
\description{Sets the default indentation for paragraphs in \code{text.view}.
Tags in the buffer may override the default.}
\usage{gtkTextViewSetIndent(object, indent)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTextView}}]  a \verb{\link{GtkTextView}}}
\item{\verb{indent}}{[integer]  indentation in pixels}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
