\alias{gtkTextViewSetLeftMargin}
\name{gtkTextViewSetLeftMargin}
\title{gtkTextViewSetLeftMargin}
\description{Sets the default left margin for text in \code{text.view}.
Tags in the buffer may override the default.}
\usage{gtkTextViewSetLeftMargin(object, left.margin)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTextView}}]  a \verb{\link{GtkTextView}}}
\item{\verb{left.margin}}{[integer]  left margin in pixels}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
