\alias{gtkTextViewSetPixelsAboveLines}
\name{gtkTextViewSetPixelsAboveLines}
\title{gtkTextViewSetPixelsAboveLines}
\description{Sets the default number of blank pixels above paragraphs in \code{text.view}.
Tags in the buffer for \code{text.view} may override the defaults.}
\usage{gtkTextViewSetPixelsAboveLines(object, pixels.above.lines)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTextView}}]  a \code{\link{GtkTextView}}}
\item{\code{pixels.above.lines}}{[integer]  pixels above paragraphs}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
