\alias{gtkTextViewSetPixelsBelowLines}
\name{gtkTextViewSetPixelsBelowLines}
\title{gtkTextViewSetPixelsBelowLines}
\description{Sets the default number of pixels of blank space
to put below paragraphs in \code{text.view}. May be overridden
by tags applied to \code{text.view}'s buffer.}
\usage{gtkTextViewSetPixelsBelowLines(object, pixels.below.lines)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTextView}}]  a \code{\link{GtkTextView}}}
\item{\code{pixels.below.lines}}{[integer]  pixels below paragraphs }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
