\alias{gtkTextViewSetPixelsInsideWrap}
\name{gtkTextViewSetPixelsInsideWrap}
\title{gtkTextViewSetPixelsInsideWrap}
\description{Sets the default number of pixels of blank space to leave between
display/wrapped lines within a paragraph. May be overridden by
tags in \code{text.view}'s buffer.}
\usage{gtkTextViewSetPixelsInsideWrap(object, pixels.inside.wrap)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTextView}}]  a \code{\link{GtkTextView}}}
\item{\code{pixels.inside.wrap}}{[integer]  default number of pixels between wrapped lines}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
