\alias{gtkTextViewSetRightMargin}
\name{gtkTextViewSetRightMargin}
\title{gtkTextViewSetRightMargin}
\description{Sets the default right margin for text in the text view.
Tags in the buffer may override the default.}
\usage{gtkTextViewSetRightMargin(object, right.margin)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTextView}}]  a \verb{\link{GtkTextView}}}
\item{\verb{right.margin}}{[integer]  right margin in pixels}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
