\alias{gtkTextViewSetWrapMode}
\name{gtkTextViewSetWrapMode}
\title{gtkTextViewSetWrapMode}
\description{Sets the line wrapping for the view.}
\usage{gtkTextViewSetWrapMode(object, wrap.mode)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTextView}}]  a \code{\link{GtkTextView}}}
\item{\code{wrap.mode}}{[\code{\link{GtkWrapMode}}]  a \code{\link{GtkWrapMode}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
