\alias{gtkTipsQueryNew}
\name{gtkTipsQueryNew}
\title{gtkTipsQueryNew}
\description{
Creates a new \code{\link{GtkTipsQuery}}.
\strong{WARNING: \code{gtk_tips_query_new} is deprecated and should not be used in newly-written code.}
}
\usage{gtkTipsQueryNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkTipsQuery}}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
