\alias{gtkTipsQueryStopQuery}
\name{gtkTipsQueryStopQuery}
\title{gtkTipsQueryStopQuery}
\description{
Stops a query.
\strong{WARNING: \code{gtk_tips_query_stop_query} is deprecated and should not be used in newly-written code.}
}
\usage{gtkTipsQueryStopQuery(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTipsQuery}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
