\alias{gtkToggleActionGetActive}
\name{gtkToggleActionGetActive}
\title{gtkToggleActionGetActive}
\description{Returns the checked state of the toggle action.}
\usage{gtkToggleActionGetActive(object)}
\arguments{\item{\verb{object}}{the action object}}
\details{Since 2.4}
\value{[logical] the checked state of the toggle action}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
