\alias{gtkToggleActionGetDrawAsRadio}
\name{gtkToggleActionGetDrawAsRadio}
\title{gtkToggleActionGetDrawAsRadio}
\description{Returns whether the action should have proxies like a radio action.}
\usage{gtkToggleActionGetDrawAsRadio(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkToggleAction}}]  the action object}}
\details{  Since  2.4}
\value{[logical]  whether the action should have proxies like a radio action.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
