\alias{gtkToggleActionSetActive}
\name{gtkToggleActionSetActive}
\title{gtkToggleActionSetActive}
\description{Sets the checked state on the toggle action.}
\usage{gtkToggleActionSetActive(object, is.active)}
\arguments{
\item{\code{object}}{[\code{\link{GtkToggleAction}}]  the action object}
\item{\code{is.active}}{[logical]  whether the action should be checked or not}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
