\alias{gtkToggleActionSetDrawAsRadio}
\name{gtkToggleActionSetDrawAsRadio}
\title{gtkToggleActionSetDrawAsRadio}
\description{Sets whether the action should have proxies like a radio action.}
\usage{gtkToggleActionSetDrawAsRadio(object, draw.as.radio)}
\arguments{
\item{\code{object}}{[\code{\link{GtkToggleAction}}]  the action object}
\item{\code{draw.as.radio}}{[logical]  whether the action should have proxies like a radio 
   action}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
