\alias{gtkToggleActionToggled}
\name{gtkToggleActionToggled}
\title{gtkToggleActionToggled}
\description{Emits the "toggled" signal on the toggle action.}
\usage{gtkToggleActionToggled(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkToggleAction}}]  the action object}}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
