\alias{gtkToggleButtonGetActive}
\name{gtkToggleButtonGetActive}
\title{gtkToggleButtonGetActive}
\description{Queries a \code{\link{GtkToggleButton}} and returns its current state. Returns \code{TRUE} if
the toggle button is pressed in and \code{FALSE} if it is raised.}
\usage{gtkToggleButtonGetActive(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkToggleButton}}] a \code{\link{GtkToggleButton}}.}}
\value{[logical] a \code{logical} value.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
