\alias{gtkToggleButtonGetMode}
\name{gtkToggleButtonGetMode}
\title{gtkToggleButtonGetMode}
\description{Retrieves whether the button is displayed as a separate indicator
and label. See \code{\link{gtkToggleButtonSetMode}}.}
\usage{gtkToggleButtonGetMode(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToggleButton}}}}
\value{[logical] \code{TRUE} if the togglebutton is drawn as a separate indicator
and label.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
