\alias{gtkToggleButtonNew}
\name{gtkToggleButtonNew}
\title{gtkToggleButtonNew}
\description{Creates a new toggle button. A widget should be packed into the button, as in \code{\link{gtkButtonNew}}.}
\usage{gtkToggleButtonNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a new toggle button.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
