\alias{gtkToggleButtonNewWithLabel}
\name{gtkToggleButtonNewWithLabel}
\title{gtkToggleButtonNewWithLabel}
\description{Creates a new toggle button with a text label.}
\usage{gtkToggleButtonNewWithLabel(label, show = TRUE)}
\arguments{\item{\verb{label}}{a string containing the message to be placed in the toggle button.}}
\value{[\code{\link{GtkWidget}}] a new toggle button.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
