\alias{gtkToggleButtonToggled}
\name{gtkToggleButtonToggled}
\title{gtkToggleButtonToggled}
\description{Emits the \code{\link{gtkToggleButtonToggled}}
signal on the \code{\link{GtkToggleButton}}. There is no good reason for an
application ever to call this function.}
\usage{gtkToggleButtonToggled(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToggleButton}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
