\alias{gtkToggleToolButtonGetActive}
\name{gtkToggleToolButtonGetActive}
\title{gtkToggleToolButtonGetActive}
\description{Queries a \code{\link{GtkToggleToolButton}} and returns its current state.
Returns \code{TRUE} if the toggle button is pressed in and \code{FALSE} if it is raised.}
\usage{gtkToggleToolButtonGetActive(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToggleToolButton}}}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if the toggle tool button is pressed in, \code{FALSE} if not}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
