\alias{gtkToggleToolButtonNew}
\name{gtkToggleToolButtonNew}
\title{gtkToggleToolButtonNew}
\description{Returns a new \verb{\link{GtkToggleToolButton}}}
\usage{gtkToggleToolButtonNew(show = TRUE)}
\details{  Since  2.4}
\value{[\verb{\link{GtkToolItem}}]  a newly created \verb{\link{GtkToggleToolButton}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
