\alias{gtkToggleToolButtonSetActive}
\name{gtkToggleToolButtonSetActive}
\title{gtkToggleToolButtonSetActive}
\description{Sets the status of the toggle tool button. Set to \code{TRUE} if you
want the GtkToggleButton to be 'pressed in', and \code{FALSE} to raise it.
This action causes the toggled signal to be emitted.}
\usage{gtkToggleToolButtonSetActive(object, is.active)}
\arguments{
\item{\code{object}}{[\code{\link{GtkToggleToolButton}}]  a \code{\link{GtkToggleToolButton}}}
\item{\code{is.active}}{[logical]  whether \code{button} should be active}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
