\alias{gtkToolbarGetNItems}
\name{gtkToolbarGetNItems}
\title{gtkToolbarGetNItems}
\description{Returns the number of items on the toolbar.}
\usage{gtkToolbarGetNItems(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolbar}}}}
\details{Since 2.4}
\value{[integer] the number of items on the toolbar}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
