\alias{gtkToolbarGetShowArrow}
\name{gtkToolbarGetShowArrow}
\title{gtkToolbarGetShowArrow}
\description{Returns whether the toolbar has an overflow menu.
See \code{\link{gtkToolbarSetShowArrow}}.}
\usage{gtkToolbarGetShowArrow(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkToolbar}}]  a \verb{\link{GtkToolbar}}}}
\details{  Since  2.4}
\value{[logical] \code{TRUE} if the toolbar has an overflow menu.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
