\alias{gtkToolbarGetStyle}
\name{gtkToolbarGetStyle}
\title{gtkToolbarGetStyle}
\description{Retrieves whether the toolbar has text, icons, or both . See
\code{\link{gtkToolbarSetStyle}}.}
\usage{gtkToolbarGetStyle(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkToolbar}}]  a \verb{\link{GtkToolbar}}}}
\value{[\code{\link{GtkToolbarStyle}}]  the current style of \code{toolbar}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
