\alias{gtkToolbarGetTooltips}
\name{gtkToolbarGetTooltips}
\title{gtkToolbarGetTooltips}
\description{Retrieves whether tooltips are enabled. See
\code{\link{gtkToolbarSetTooltips}}.}
\usage{gtkToolbarGetTooltips(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkToolbar}}]  a \verb{\link{GtkToolbar}}}}
\value{[logical] \code{TRUE} if tooltips are enabled}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
