\alias{gtkToolbarInsert}
\name{gtkToolbarInsert}
\title{gtkToolbarInsert}
\description{Insert a \verb{\link{GtkToolItem}} into the toolbar at position \code{pos}. If \code{pos} is
0 the item is prepended to the start of the toolbar. If \code{pos} is
negative, the item is appended to the end of the toolbar.}
\usage{gtkToolbarInsert(object, item, pos)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkToolbar}}]  a \verb{\link{GtkToolbar}}}
\item{\verb{item}}{[\verb{\link{GtkToolItem}}]  a \verb{\link{GtkToolItem}}}
\item{\verb{pos}}{[integer]  the position of the new item}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
