\alias{gtkToolbarInsertSpace}
\name{gtkToolbarInsertSpace}
\title{gtkToolbarInsertSpace}
\description{
Inserts a new space in the toolbar at the specified position.
\strong{WARNING: \code{gtk_toolbar_insert_space} is deprecated and should not be used in newly-written code.  
Inserts a new space in the toolbar at the specified position.  }
}
\usage{gtkToolbarInsertSpace(object, position)}
\arguments{
\item{\code{object}}{[\code{\link{GtkToolbar}}]  a \code{\link{GtkToolbar}}}
\item{\code{position}}{[integer]  the number of widgets after which a space should be inserted.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
