\alias{gtkToolbarPrependSpace}
\name{gtkToolbarPrependSpace}
\title{gtkToolbarPrependSpace}
\description{
Adds a new space to the beginning of the toolbar.
\strong{WARNING: \code{gtk_toolbar_prepend_space} is deprecated and should not be used in newly-written code.}
}
\usage{gtkToolbarPrependSpace(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkToolbar}}]  a \verb{\link{GtkToolbar}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
