\alias{gtkToolbarRemoveSpace}
\name{gtkToolbarRemoveSpace}
\title{gtkToolbarRemoveSpace}
\description{
Removes a space from the specified position.
\strong{WARNING: \code{gtk_toolbar_remove_space} is deprecated and should not be used in newly-written code.}
}
\usage{gtkToolbarRemoveSpace(object, position)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkToolbar}}]  a \verb{\link{GtkToolbar}}.}
\item{\verb{position}}{[integer]  the index of the space to remove.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
