\alias{gtkToolbarSetOrientation}
\name{gtkToolbarSetOrientation}
\title{gtkToolbarSetOrientation}
\description{Sets whether a toolbar should appear horizontally or vertically.}
\usage{gtkToolbarSetOrientation(object, orientation)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkToolbar}}]  a \verb{\link{GtkToolbar}}.}
\item{\verb{orientation}}{[\code{\link{GtkOrientation}}]  a new \code{\link{GtkOrientation}}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
