\alias{gtkToolButtonGetIconName}
\name{gtkToolButtonGetIconName}
\title{gtkToolButtonGetIconName}
\description{Returns the name of the themed icon for the tool button,
see \code{\link{gtkToolButtonSetIconName}}.}
\usage{gtkToolButtonGetIconName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolButton}}}}
\details{Since 2.8}
\value{[character] the icon name or \code{NULL} if the tool button has
no themed icon}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
