\alias{gtkToolButtonGetLabel}
\name{gtkToolButtonGetLabel}
\title{gtkToolButtonGetLabel}
\description{Returns the label used by the tool button, or \code{NULL} if the tool button
doesn't have a label. or uses a the label from a stock item. and must not be modified or freed.}
\usage{gtkToolButtonGetLabel(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolButton}}}}
\details{Since 2.4}
\value{[character] The label, or \code{NULL}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
