\alias{gtkToolButtonGetLabelWidget}
\name{gtkToolButtonGetLabelWidget}
\title{gtkToolButtonGetLabelWidget}
\description{Returns the widget used as label on \code{button}. See
\code{\link{gtkToolButtonSetLabelWidget}}.}
\usage{gtkToolButtonGetLabelWidget(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolButton}}}}
\details{Since 2.4}
\value{[\code{\link{GtkWidget}}] The widget used as label on \code{button}, or \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
