\alias{gtkToolButtonGetStockId}
\name{gtkToolButtonGetStockId}
\title{gtkToolButtonGetStockId}
\description{Returns the name of the stock item. See \code{\link{gtkToolButtonSetStockId}}.}
\usage{gtkToolButtonGetStockId(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolButton}}}}
\details{Since 2.4}
\value{[character] the name of the stock item for \code{button}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
