\alias{gtkToolButtonGetUseUnderline}
\name{gtkToolButtonGetUseUnderline}
\title{gtkToolButtonGetUseUnderline}
\description{Returns whether underscores in the label property are used as mnemonics
on menu items on the overflow menu. See \code{\link{gtkToolButtonSetUseUnderline}}.}
\usage{gtkToolButtonGetUseUnderline(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolButton}}}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if underscores in the label property are used as
mnemonics on menu items on the overflow menu.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
