\alias{gtkToolButtonNew}
\name{gtkToolButtonNew}
\title{gtkToolButtonNew}
\description{Creates a new \verb{\link{GtkToolButton}} using \code{icon.widget} as icon and \code{label} as
label.}
\usage{gtkToolButtonNew(icon.widget = NULL, label = NULL, show = TRUE)}
\arguments{
\item{\verb{icon.widget}}{[\verb{\link{GtkWidget}}]  a widget that will be used as icon widget, or \code{NULL}}
\item{\verb{label}}{[character]  a string that will be used as label, or \code{NULL}}
}
\details{  Since  2.4}
\value{[\verb{\link{GtkToolItem}}]  A new \verb{\link{GtkToolButton}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
