\alias{gtkToolButtonSetIconName}
\name{gtkToolButtonSetIconName}
\title{gtkToolButtonSetIconName}
\description{Sets the icon for the tool button from a named themed icon.
See the docs for \code{\link{GtkIconTheme}} for more details.
The "icon\_name" property only has an effect if not
overridden by non-\code{NULL} "label", "icon\_widget" and "stock\_id"
properties.}
\usage{gtkToolButtonSetIconName(object, icon.name)}
\arguments{
\item{\code{object}}{[\code{\link{GtkToolButton}}]  a \code{\link{GtkToolButton}}}
\item{\code{icon.name}}{[character]  the name of the themed icon}
}
\details{  Since  2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
