\alias{gtkToolButtonSetStockId}
\name{gtkToolButtonSetStockId}
\title{gtkToolButtonSetStockId}
\description{Sets the name of the stock item. See \code{\link{gtkToolButtonNewFromStock}}.
The stock\_id property only has an effect if not
overridden by non-\code{NULL} "label" and "icon\_widget" properties.}
\usage{gtkToolButtonSetStockId(object, stock.id = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkToolButton}}]  a \code{\link{GtkToolButton}}}
\item{\code{stock.id}}{[character]  a name of a stock item, or \code{NULL}}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
