\alias{gtkToolItemGetExpand}
\name{gtkToolItemGetExpand}
\title{gtkToolItemGetExpand}
\description{Returns whether \code{tool.item} is allocated extra space.
See \code{\link{gtkToolItemSetExpand}}.}
\usage{gtkToolItemGetExpand(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkToolItem}}]  a \code{\link{gtkToolItem}} }}
\details{  Since  2.4}
\value{[logical]  \code{TRUE} if \code{tool.item} is allocated extra space.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
