\alias{gtkToolItemGetHomogeneous}
\name{gtkToolItemGetHomogeneous}
\title{gtkToolItemGetHomogeneous}
\description{Returns whether \code{tool.item} is the same size as other homogeneous
items. See \code{\link{gtkToolItemSetHomogeneous}}.}
\usage{gtkToolItemGetHomogeneous(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolItem}}}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if the item is the same size as other homogeneous
items.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
