\alias{gtkToolItemGetIconSize}
\name{gtkToolItemGetIconSize}
\title{gtkToolItemGetIconSize}
\description{Returns the icon size used for \code{tool.item}. Custom subclasses of
\verb{\link{GtkToolItem}} should call this function to find out what size icons
they should use.}
\usage{gtkToolItemGetIconSize(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkToolItem}}]  a \code{\link{gtkToolItem}}}}
\details{  Since  2.4}
\value{[\code{\link{GtkIconSize}}]  a \code{\link{GtkIconSize}} indicating the icon size used for \code{tool.item}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
