\alias{gtkToolItemGetIsImportant}
\name{gtkToolItemGetIsImportant}
\title{gtkToolItemGetIsImportant}
\description{Returns whether \code{tool.item} is considered important. See
\code{\link{gtkToolItemSetIsImportant}}}
\usage{gtkToolItemGetIsImportant(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolItem}}}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if \code{tool.item} is considered important.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
