\alias{gtkToolItemGetOrientation}
\name{gtkToolItemGetOrientation}
\title{gtkToolItemGetOrientation}
\description{Returns the orientation used for \code{tool.item}. Custom subclasses of
\code{\link{GtkToolItem}} should call this function to find out what size icons
they should use.}
\usage{gtkToolItemGetOrientation(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolItem}}}}
\details{Since 2.4}
\value{[\code{\link{GtkOrientation}}] a \code{\link{GtkOrientation}} indicating the orientation
used for \code{tool.item}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
