\alias{gtkToolItemGetTextAlignment}
\name{gtkToolItemGetTextAlignment}
\title{gtkToolItemGetTextAlignment}
\description{Returns the text alignment used for \code{tool.item}. Custom subclasses of
\code{\link{GtkToolItem}} should call this function to find out how text should
be aligned.}
\usage{gtkToolItemGetTextAlignment(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolItem}}:}}
\details{Since 2.20}
\value{[numeric] a \verb{numeric} indicating the horizontal text alignment
used for \code{tool.item}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
