\alias{gtkToolItemGetTextOrientation}
\name{gtkToolItemGetTextOrientation}
\title{gtkToolItemGetTextOrientation}
\description{Returns the text orientation used for \code{tool.item}. Custom subclasses of
\code{\link{GtkToolItem}} should call this function to find out how text should
be orientated.}
\usage{gtkToolItemGetTextOrientation(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolItem}}}}
\details{Since 2.20}
\value{[\code{\link{GtkOrientation}}] a \code{\link{GtkOrientation}} indicating the text orientation
used for \code{tool.item}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
