\alias{gtkToolItemGetTextSizeGroup}
\name{gtkToolItemGetTextSizeGroup}
\title{gtkToolItemGetTextSizeGroup}
\description{Returns the size group used for labels in \code{tool.item}. Custom subclasses of
\code{\link{GtkToolItem}} should call this function and use the size group for labels.}
\usage{gtkToolItemGetTextSizeGroup(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolItem}}}}
\details{Since 2.20}
\value{[\code{\link{GtkSizeGroup}}] a \code{\link{GtkSizeGroup}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
