\alias{gtkToolItemGetUseDragWindow}
\name{gtkToolItemGetUseDragWindow}
\title{gtkToolItemGetUseDragWindow}
\description{Returns whether \code{toolitem} has a drag window. See
\code{\link{gtkToolItemSetUseDragWindow}}.}
\usage{gtkToolItemGetUseDragWindow(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkToolItem}}]  a \code{\link{GtkToolItem}} }}
\details{  Since  2.4}
\value{[logical]  \code{TRUE} if \code{toolitem} uses a drag window.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
