\alias{gtkToolItemGetVisibleHorizontal}
\name{gtkToolItemGetVisibleHorizontal}
\title{gtkToolItemGetVisibleHorizontal}
\description{Returns whether the \code{tool.item} is visible on toolbars that are
docked horizontally.}
\usage{gtkToolItemGetVisibleHorizontal(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolItem}}}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if \code{tool.item} is visible on toolbars that are
docked horizontally.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
