\alias{gtkToolItemGetVisibleVertical}
\name{gtkToolItemGetVisibleVertical}
\title{gtkToolItemGetVisibleVertical}
\description{Returns whether \code{tool.item} is visible when the toolbar is docked vertically.
See \code{\link{gtkToolItemSetVisibleVertical}}.}
\usage{gtkToolItemGetVisibleVertical(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolItem}}}}
\details{Since 2.4}
\value{[logical] Whether \code{tool.item} is visible when the toolbar is docked vertically}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
