\alias{gtkToolItemGroupGetCollapsed}
\name{gtkToolItemGroupGetCollapsed}
\title{gtkToolItemGroupGetCollapsed}
\description{Gets whether \code{group} is collapsed or expanded.}
\usage{gtkToolItemGroupGetCollapsed(object)}
\arguments{\item{\verb{object}}{a GtkToolItemGroup}}
\details{Since 2.20}
\value{[logical] \code{TRUE} if \code{group} is collapsed, \code{FALSE} if it is expanded}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
