\alias{gtkToolItemGroupGetDropItem}
\name{gtkToolItemGroupGetDropItem}
\title{gtkToolItemGroupGetDropItem}
\description{Gets the tool item at position (x, y).}
\usage{gtkToolItemGroupGetDropItem(object, x, y)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolItemGroup}}}
\item{\verb{x}}{the x position}
\item{\verb{y}}{the y position}
}
\details{Since 2.20}
\value{[\code{\link{GtkToolItem}}] the \code{\link{GtkToolItem}} at position (x, y)}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
