\alias{gtkToolItemGroupGetEllipsize}
\name{gtkToolItemGroupGetEllipsize}
\title{gtkToolItemGroupGetEllipsize}
\description{Gets the ellipsization mode of \code{group}.}
\usage{gtkToolItemGroupGetEllipsize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolItemGroup}}}}
\details{Since 2.20}
\value{[\code{\link{PangoEllipsizeMode}}] the \code{\link{PangoEllipsizeMode}} of \code{group}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
