\alias{gtkToolItemGroupGetHeaderRelief}
\name{gtkToolItemGroupGetHeaderRelief}
\title{gtkToolItemGroupGetHeaderRelief}
\description{Gets the relief mode of the header button of \code{group}.}
\usage{gtkToolItemGroupGetHeaderRelief(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolItemGroup}}}}
\details{Since 2.20}
\value{[\code{\link{GtkReliefStyle}}] the \code{\link{GtkReliefStyle}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
