\alias{gtkToolItemGroupGetItemPosition}
\name{gtkToolItemGroupGetItemPosition}
\title{gtkToolItemGroupGetItemPosition}
\description{Gets the position of \code{item} in \code{group} as index.}
\usage{gtkToolItemGroupGetItemPosition(object, item)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolItemGroup}}}
\item{\verb{item}}{a \code{\link{GtkToolItem}}}
}
\details{Since 2.20}
\value{[integer] the index of \code{item} in \code{group} or -1 if \code{item} is no child of \code{group}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
