\alias{gtkToolItemGroupGetLabel}
\name{gtkToolItemGroupGetLabel}
\title{gtkToolItemGroupGetLabel}
\description{Gets the label of \code{group}.}
\usage{gtkToolItemGroupGetLabel(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolItemGroup}}}}
\details{Since 2.20}
\value{[character] the label of \code{group}. The label is an internal string of \code{group}
and must not be modified. Note that \code{NULL} is returned if a custom
label has been set with \code{\link{gtkToolItemGroupSetLabelWidget}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
