\alias{gtkToolItemGroupGetLabelWidget}
\name{gtkToolItemGroupGetLabelWidget}
\title{gtkToolItemGroupGetLabelWidget}
\description{Gets the label widget of \code{group}.
See \code{\link{gtkToolItemGroupSetLabelWidget}}.}
\usage{gtkToolItemGroupGetLabelWidget(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolItemGroup}}}}
\details{Since 2.20}
\value{[\code{\link{GtkWidget}}] the label widget of \code{group}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
