\alias{gtkToolItemGroupGetNItems}
\name{gtkToolItemGroupGetNItems}
\title{gtkToolItemGroupGetNItems}
\description{Gets the number of tool items in \code{group}.}
\usage{gtkToolItemGroupGetNItems(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolItemGroup}}}}
\details{Since 2.20}
\value{[numeric] the number of tool items in \code{group}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
