\alias{gtkToolItemGroupGetNthItem}
\name{gtkToolItemGroupGetNthItem}
\title{gtkToolItemGroupGetNthItem}
\description{Gets the tool item at \code{index} in group.}
\usage{gtkToolItemGroupGetNthItem(object, index)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolItemGroup}}}
\item{\verb{index}}{the index}
}
\details{Since 2.20}
\value{[\code{\link{GtkToolItem}}] the \code{\link{GtkToolItem}} at index}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
