\alias{gtkToolItemGroupInsert}
\name{gtkToolItemGroupInsert}
\title{gtkToolItemGroupInsert}
\description{Inserts \code{item} at \code{position} in the list of children of \code{group}.}
\usage{gtkToolItemGroupInsert(object, item, position)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolItemGroup}}}
\item{\verb{item}}{the \code{\link{GtkToolItem}} to insert into \code{group}}
\item{\verb{position}}{the position of \code{item} in \code{group}, starting with 0.
The position -1 means end of list.}
}
\details{Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
