\alias{gtkToolItemGroupNew}
\name{gtkToolItemGroupNew}
\title{gtkToolItemGroupNew}
\description{Creates a new tool item group with label \code{label}.}
\usage{gtkToolItemGroupNew(label, show = TRUE)}
\arguments{\item{\verb{label}}{the label of the new group}}
\details{Since 2.20}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkToolItemGroup}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
