\alias{gtkToolItemGroupSetCollapsed}
\name{gtkToolItemGroupSetCollapsed}
\title{gtkToolItemGroupSetCollapsed}
\description{Sets whether the \code{group} should be collapsed or expanded.}
\usage{gtkToolItemGroupSetCollapsed(object, collapsed)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolItemGroup}}}
\item{\verb{collapsed}}{whether the \code{group} should be collapsed or expanded}
}
\details{Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
