\alias{gtkToolItemGroupSetEllipsize}
\name{gtkToolItemGroupSetEllipsize}
\title{gtkToolItemGroupSetEllipsize}
\description{Sets the ellipsization mode which should be used by labels in \code{group}.}
\usage{gtkToolItemGroupSetEllipsize(object, ellipsize)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolItemGroup}}}
\item{\verb{ellipsize}}{the \code{\link{PangoEllipsizeMode}} labels in \code{group} should use}
}
\details{Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
