\alias{gtkToolItemGroupSetHeaderRelief}
\name{gtkToolItemGroupSetHeaderRelief}
\title{gtkToolItemGroupSetHeaderRelief}
\description{Set the button relief of the group header.
See \code{\link{gtkButtonSetRelief}} for details.}
\usage{gtkToolItemGroupSetHeaderRelief(object, style)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolItemGroup}}}
\item{\verb{style}}{the \code{\link{GtkReliefStyle}}}
}
\details{Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
