\alias{gtkToolItemGroupSetLabel}
\name{gtkToolItemGroupSetLabel}
\title{gtkToolItemGroupSetLabel}
\description{Sets the label of the tool item group. The label is displayed in the header
of the group.}
\usage{gtkToolItemGroupSetLabel(object, label)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolItemGroup}}}
\item{\verb{label}}{the new human-readable label of of the group}
}
\details{Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
