\alias{gtkToolItemGroupSetLabelWidget}
\name{gtkToolItemGroupSetLabelWidget}
\title{gtkToolItemGroupSetLabelWidget}
\description{Sets the label of the tool item group.
The label widget is displayed in the header of the group, in place
of the usual label.}
\usage{gtkToolItemGroupSetLabelWidget(object, label.widget)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolItemGroup}}}
\item{\verb{label.widget}}{the widget to be displayed in place of the usual label}
}
\details{Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
