\alias{gtkToolItemNew}
\name{gtkToolItemNew}
\title{gtkToolItemNew}
\description{Creates a new \verb{\link{GtkToolItem}}}
\usage{gtkToolItemNew(show = TRUE)}
\details{  Since  2.4}
\value{[\verb{\link{GtkToolItem}}]  the new \verb{\link{GtkToolItem}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
