\alias{gtkToolItemRetrieveProxyMenuItem}
\name{gtkToolItemRetrieveProxyMenuItem}
\title{gtkToolItemRetrieveProxyMenuItem}
\description{Returns the \verb{\link{GtkMenuItem}} that was last set by
\code{\link{gtkToolItemSetProxyMenuItem}}, ie. the \verb{\link{GtkMenuItem}}
that is going to appear in the overflow menu.}
\usage{gtkToolItemRetrieveProxyMenuItem(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkToolItem}}]  a \code{\link{gtkToolItem}} }}
\details{  Since  2.4}
\value{[\verb{\link{GtkWidget}}]  The \verb{\link{GtkMenuItem}} that is going to appear in the
overflow menu for \code{tool.item}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
