\alias{gtkToolItemSetExpand}
\name{gtkToolItemSetExpand}
\title{gtkToolItemSetExpand}
\description{Sets whether \code{tool.item} is allocated extra space when there
is more room on the toolbar then needed for the items. The
effect is that the item gets bigger when the toolbar gets bigger
and smaller when the toolbar gets smaller.}
\usage{gtkToolItemSetExpand(object, expand)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolItem}}}
\item{\verb{expand}}{Whether \code{tool.item} is allocated extra space}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
