\alias{gtkToolItemSetTooltipMarkup}
\name{gtkToolItemSetTooltipMarkup}
\title{gtkToolItemSetTooltipMarkup}
\description{Sets the markup text to be displayed as tooltip on the item.
See \code{\link{gtkWidgetSetTooltipMarkup}}.}
\usage{gtkToolItemSetTooltipMarkup(object, markup)}
\arguments{
\item{\code{object}}{[\code{\link{GtkToolItem}}]  a \code{\link{gtkToolItem}} }
\item{\code{markup}}{[character]  markup text to be used as tooltip for \code{tool.item}}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
