\alias{gtkToolItemSetVisibleVertical}
\name{gtkToolItemSetVisibleVertical}
\title{gtkToolItemSetVisibleVertical}
\description{Sets whether \code{tool.item} is visible when the toolbar is docked
vertically. Some tool items, such as text entries, are too wide to be
useful on a vertically docked toolbar. If \code{visible.vertical} is \code{FALSE}
\code{tool.item} will not appear on toolbars that are docked vertically.}
\usage{gtkToolItemSetVisibleVertical(object, visible.vertical)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolItem}}}
\item{\verb{visible.vertical}}{whether \code{tool.item} is visible when the toolbar
is in vertical mode}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
