\alias{gtkToolItemToolbarReconfigured}
\name{gtkToolItemToolbarReconfigured}
\title{gtkToolItemToolbarReconfigured}
\description{Emits the signal \code{\link{gtkToolItemToolbarReconfigured}} on \code{tool.item}.
\code{\link{GtkToolbar}} and other \code{\link{GtkToolShell}} implementations use this function
to notify children, when some aspect of their configuration changes.}
\usage{gtkToolItemToolbarReconfigured(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolItem}}}}
\details{Since 2.14}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
