\alias{gtkToolPaletteGetDragItem}
\name{gtkToolPaletteGetDragItem}
\title{gtkToolPaletteGetDragItem}
\description{Get the dragged item from the selection.
This could be a \code{\link{GtkToolItem}} or a \code{\link{GtkToolItemGroup}}.}
\usage{gtkToolPaletteGetDragItem(object, selection)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolPalette}}}
\item{\verb{selection}}{a \code{\link{GtkSelectionData}}}
}
\details{Since 2.20}
\value{[\code{\link{GtkWidget}}] the dragged item in selection}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
