\alias{gtkToolPaletteGetDragTargetGroup}
\name{gtkToolPaletteGetDragTargetGroup}
\title{gtkToolPaletteGetDragTargetGroup}
\description{Get the target entry for a dragged \code{\link{GtkToolItemGroup}}.}
\usage{gtkToolPaletteGetDragTargetGroup()}
\details{Since 2.20}
\value{[\code{\link{GtkTargetEntry}}] the \code{\link{GtkTargetEntry}} for a dragged group}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
