\alias{gtkToolPaletteGetDragTargetItem}
\name{gtkToolPaletteGetDragTargetItem}
\title{gtkToolPaletteGetDragTargetItem}
\description{Gets the target entry for a dragged \code{\link{GtkToolItem}}.}
\usage{gtkToolPaletteGetDragTargetItem()}
\details{Since 2.20}
\value{[\code{\link{GtkTargetEntry}}] the \code{\link{GtkTargetEntry}} for a dragged item.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
