\alias{gtkToolPaletteGetDropGroup}
\name{gtkToolPaletteGetDropGroup}
\title{gtkToolPaletteGetDropGroup}
\description{Gets the group at position (x, y).}
\usage{gtkToolPaletteGetDropGroup(object, x, y)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolPalette}}}
\item{\verb{x}}{the x position}
\item{\verb{y}}{the y position}
}
\details{Since 2.20}
\value{[\code{\link{GtkToolItemGroup}}] the \code{\link{GtkToolItemGroup}} at position or \code{NULL}
if there is no such group}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
