\alias{gtkToolPaletteGetDropItem}
\name{gtkToolPaletteGetDropItem}
\title{gtkToolPaletteGetDropItem}
\description{Gets the item at position (x, y).
See \code{\link{gtkToolPaletteGetDropGroup}}.}
\usage{gtkToolPaletteGetDropItem(object, x, y)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolPalette}}}
\item{\verb{x}}{the x position}
\item{\verb{y}}{the y position}
}
\details{Since 2.20}
\value{[\code{\link{GtkToolItem}}] the \code{\link{GtkToolItem}} at position or \code{NULL} if there is no such item}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
