\alias{gtkToolPaletteGetExclusive}
\name{gtkToolPaletteGetExclusive}
\title{gtkToolPaletteGetExclusive}
\description{Gets whether \code{group} is exclusive or not.
See \code{\link{gtkToolPaletteSetExclusive}}.}
\usage{gtkToolPaletteGetExclusive(object, group)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolPalette}}}
\item{\verb{group}}{a \code{\link{GtkToolItemGroup}} which is a child of palette}
}
\details{Since 2.20}
\value{[logical] \code{TRUE} if \code{group} is exclusive}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
