\alias{gtkToolPaletteGetExpand}
\name{gtkToolPaletteGetExpand}
\title{gtkToolPaletteGetExpand}
\description{Gets whether group should be given extra space.
See \code{\link{gtkToolPaletteSetExpand}}.}
\usage{gtkToolPaletteGetExpand(object, group)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolPalette}}}
\item{\verb{group}}{a \code{\link{GtkToolItemGroup}} which is a child of palette}
}
\details{Since 2.20}
\value{[logical] \code{TRUE} if group should be given extra space, \code{FALSE} otherwise}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
