\alias{gtkToolPaletteGetHadjustment}
\name{gtkToolPaletteGetHadjustment}
\title{gtkToolPaletteGetHadjustment}
\description{Gets the horizontal adjustment of the tool palette.}
\usage{gtkToolPaletteGetHadjustment(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolPalette}}}}
\details{Since 2.20}
\value{[\code{\link{GtkAdjustment}}] the horizontal adjustment of \code{palette}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
