\alias{gtkToolPaletteGetIconSize}
\name{gtkToolPaletteGetIconSize}
\title{gtkToolPaletteGetIconSize}
\description{Gets the size of icons in the tool palette.
See \code{\link{gtkToolPaletteSetIconSize}}.}
\usage{gtkToolPaletteGetIconSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolPalette}}}}
\details{Since 2.20}
\value{[\code{\link{GtkIconSize}}] the \code{\link{GtkIconSize}} of icons in the tool palette. \emph{[  \acronym{type}   int]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
