\alias{gtkToolPaletteGetStyle}
\name{gtkToolPaletteGetStyle}
\title{gtkToolPaletteGetStyle}
\description{Gets the style (icons, text or both) of items in the tool palette.}
\usage{gtkToolPaletteGetStyle(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolPalette}}}}
\details{Since 2.20}
\value{[\code{\link{GtkToolbarStyle}}] the \code{\link{GtkToolbarStyle}} of items in the tool palette.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
