\alias{gtkToolPaletteGetVadjustment}
\name{gtkToolPaletteGetVadjustment}
\title{gtkToolPaletteGetVadjustment}
\description{Gets the vertical adjustment of the tool palette.}
\usage{gtkToolPaletteGetVadjustment(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolPalette}}}}
\details{Since 2.20}
\value{[\code{\link{GtkAdjustment}}] the vertical adjustment of \code{palette}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
