\alias{gtkToolPaletteNew}
\name{gtkToolPaletteNew}
\title{gtkToolPaletteNew}
\description{Creates a new tool palette.}
\usage{gtkToolPaletteNew(show = TRUE)}
\details{Since 2.20}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkToolPalette}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
