\alias{gtkToolPaletteSetDragSource}
\name{gtkToolPaletteSetDragSource}
\title{gtkToolPaletteSetDragSource}
\description{Sets the tool palette as a drag source.
Enables all groups and items in the tool palette as drag sources
on button 1 and button 3 press with copy and move actions.
See \code{\link{gtkDragSourceSet}}.}
\usage{gtkToolPaletteSetDragSource(object, targets)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolPalette}}}
\item{\verb{targets}}{the \verb{GtkToolPaletteDragTarget}s
which the widget should support}
}
\details{Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
