\alias{gtkToolPaletteSetExclusive}
\name{gtkToolPaletteSetExclusive}
\title{gtkToolPaletteSetExclusive}
\description{Sets whether the group should be exclusive or not.
If an exclusive group is expanded all other groups are collapsed.}
\usage{gtkToolPaletteSetExclusive(object, group, exclusive)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolPalette}}}
\item{\verb{group}}{a \code{\link{GtkToolItemGroup}} which is a child of palette}
\item{\verb{exclusive}}{whether the group should be exclusive or not}
}
\details{Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
