\alias{gtkToolPaletteSetExpand}
\name{gtkToolPaletteSetExpand}
\title{gtkToolPaletteSetExpand}
\description{Sets whether the group should be given extra space.}
\usage{gtkToolPaletteSetExpand(object, group, expand)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolPalette}}}
\item{\verb{group}}{a \code{\link{GtkToolItemGroup}} which is a child of palette}
\item{\verb{expand}}{whether the group should be given extra space}
}
\details{Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
