\alias{gtkToolPaletteSetGroupPosition}
\name{gtkToolPaletteSetGroupPosition}
\title{gtkToolPaletteSetGroupPosition}
\description{Sets the position of the group as an index of the tool palette.
If position is 0 the group will become the first child, if position is
-1 it will become the last child.}
\usage{gtkToolPaletteSetGroupPosition(object, group, position)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolPalette}}}
\item{\verb{group}}{a \code{\link{GtkToolItemGroup}} which is a child of palette}
\item{\verb{position}}{a new index for group}
}
\details{Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
