\alias{gtkToolPaletteSetIconSize}
\name{gtkToolPaletteSetIconSize}
\title{gtkToolPaletteSetIconSize}
\description{Sets the size of icons in the tool palette.}
\usage{gtkToolPaletteSetIconSize(object, icon.size)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolPalette}}}
\item{\verb{icon.size}}{the \code{\link{GtkIconSize}} that icons in the tool
palette shall have. \emph{[  \acronym{type}   int]}}
}
\details{Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
