\alias{gtkToolPaletteSetStyle}
\name{gtkToolPaletteSetStyle}
\title{gtkToolPaletteSetStyle}
\description{Sets the style (text, icons or both) of items in the tool palette.}
\usage{gtkToolPaletteSetStyle(object, style)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolPalette}}}
\item{\verb{style}}{the \code{\link{GtkToolbarStyle}} that items in the tool palette shall have}
}
\details{Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
