\alias{gtkToolPaletteUnsetIconSize}
\name{gtkToolPaletteUnsetIconSize}
\title{gtkToolPaletteUnsetIconSize}
\description{Unsets the tool palette icon size set with \code{\link{gtkToolPaletteSetIconSize}},
so that user preferences will be used to determine the icon size.}
\usage{gtkToolPaletteUnsetIconSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolPalette}}}}
\details{Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
