\alias{gtkToolPaletteUnsetStyle}
\name{gtkToolPaletteUnsetStyle}
\title{gtkToolPaletteUnsetStyle}
\description{Unsets a toolbar style set with \code{\link{gtkToolPaletteSetStyle}},
so that user preferences will be used to determine the toolbar style.}
\usage{gtkToolPaletteUnsetStyle(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolPalette}}}}
\details{Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
