\alias{gtkToolShellGetTextAlignment}
\name{gtkToolShellGetTextAlignment}
\title{gtkToolShellGetTextAlignment}
\description{Retrieves the current text alignment for the tool shell. Tool items must not
call this function directly, but rely on \code{\link{gtkToolItemGetTextAlignment}}
instead.}
\usage{gtkToolShellGetTextAlignment(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolShell}}}}
\details{Since 2.14}
\value{[numeric] the current text alignment of \code{shell}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
