\alias{gtkTooltipsDisable}
\name{gtkTooltipsDisable}
\title{gtkTooltipsDisable}
\description{
Causes all tooltips in \code{tooltips} to become inactive. Any widgets that have tips associated with that group will no longer display their tips until they are enabled again with \code{\link{gtkTooltipsEnable}}.
\strong{WARNING: \code{gtk_tooltips_disable} has been deprecated since version 2.12 and should not be used in newly-written code. }
}
\usage{gtkTooltipsDisable(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTooltips}}] a \code{\link{GtkTooltips}}.}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
