\alias{gtkTooltipsEnable}
\name{gtkTooltipsEnable}
\title{gtkTooltipsEnable}
\description{Allows the user to see your tooltips as they navigate your application.}
\usage{gtkTooltipsEnable(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTooltips}}] a \code{\link{GtkTooltips}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
