\alias{gtkTooltipSetIcon}
\name{gtkTooltipSetIcon}
\title{gtkTooltipSetIcon}
\description{Sets the icon of the tooltip (which is in front of the text) to be
\code{pixbuf}.  If \code{pixbuf} is \code{NULL}, the image will be hidden.}
\usage{gtkTooltipSetIcon(object, pixbuf)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTooltip}}]  a \code{\link{GtkTooltip}}}
\item{\code{pixbuf}}{[\code{\link{GdkPixbuf}}]  a \code{\link{GdkPixbuf}}, or \code{NULL}}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
