\alias{gtkTooltipsetMarkup}
\name{gtkTooltipsetMarkup}
\title{gtkTooltipsetMarkup}
\description{Sets the text of the tooltip to be \code{markup}, which is marked up
with the Pango text markup language.
If \code{markup} is \code{NULL}, the label will be hidden.}
\usage{gtkTooltipsetMarkup(object, markup)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTooltip}}]  a \code{\link{GtkTooltip}}}
\item{\code{markup}}{[character]  a markup string (see Pango markup format) or \code{NULL}}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
