\alias{gtkTooltipsetText}
\name{gtkTooltipsetText}
\title{gtkTooltipsetText}
\description{Sets the text of the tooltip to be \code{text}. If \code{text} is \code{NULL}, the label
will be hidden. See also \code{\link{gtkTooltipsetMarkup}}.}
\usage{gtkTooltipsetText(object, text)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTooltip}}]  a \code{\link{GtkTooltip}}}
\item{\code{text}}{[character]  a text string or \code{NULL}}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
