\alias{gtkTooltipsForceWindow}
\name{gtkTooltipsForceWindow}
\title{gtkTooltipsForceWindow}
\description{Ensures that the window used for displaying the given \code{tooltips} is created.}
\usage{gtkTooltipsForceWindow(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTooltips}}] a \code{GtkToolTips}}}
\details{Applications should never have to call this function, since GTK+ takes
care of this.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
