\alias{gtkTooltipsNew}
\name{gtkTooltipsNew}
\title{gtkTooltipsNew}
\description{Creates an empty group of tooltips. This function initialises a \code{\link{GtkTooltips}} structure. Without at least one such structure, you can not add tips to your application.}
\usage{gtkTooltipsNew()}
\value{[\code{\link{GtkTooltips}}] a new \code{\link{GtkTooltips}} group for you to use.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
