\alias{gtkTooltipsSetDelay}
\name{gtkTooltipsSetDelay}
\title{gtkTooltipsSetDelay}
\description{
Sets the time between the user moving the mouse over a widget and the widget's tooltip appearing.
\strong{WARNING: \code{gtk_tooltips_set_delay} is deprecated and should not be used in newly-written code.}
}
\usage{gtkTooltipsSetDelay(object, delay)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTooltips}}] a \code{\link{GtkTooltips}}.}
\item{\code{delay}}{[numeric] an integer value representing milliseconds.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
