\alias{gtkTooltipTriggerTooltipQuery}
\name{gtkTooltipTriggerTooltipQuery}
\title{gtkTooltipTriggerTooltipQuery}
\description{Triggers a new tooltip query on \code{display}, in order to update the current
visible tooltip, or to show/hide the current tooltip.  This function is
useful to call when, for example, the state of the widget changed by a
key press.}
\usage{gtkTooltipTriggerTooltipQuery(display)}
\arguments{\item{\verb{display}}{a \code{\link{GdkDisplay}}}}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
