\alias{gtkTreeModelFilterGetModel}
\name{gtkTreeModelFilterGetModel}
\title{gtkTreeModelFilterGetModel}
\description{Returns a pointer to the child model of \code{filter}.}
\usage{gtkTreeModelFilterGetModel(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeModelFilter}}]  A \verb{\link{GtkTreeModelFilter}}.}}
\details{  Since  2.4}
\value{[\verb{\link{GtkTreeModel}}]  A pointer to a \verb{\link{GtkTreeModel}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
