\alias{gtkTreeModelFilterNew}
\name{gtkTreeModelFilterNew}
\title{gtkTreeModelFilterNew}
\description{Creates a new \verb{\link{GtkTreeModel}}, with \code{child.model} as the child_model
and \code{root} as the virtual root.}
\usage{gtkTreeModelFilterNew(child.model, root = NULL)}
\arguments{
\item{\verb{child.model}}{[\verb{\link{GtkTreeModel}}]  A \verb{\link{GtkTreeModel}}.}
\item{\verb{root}}{[\verb{\link{GtkTreePath}}]  A \verb{\link{GtkTreePath}} or \code{NULL}.}
}
\details{  Since  2.4}
\value{[\verb{\link{GtkTreeModel}}]  A new \verb{\link{GtkTreeModel}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
