\alias{gtkTreeModelFilterRefilter}
\name{gtkTreeModelFilterRefilter}
\title{gtkTreeModelFilterRefilter}
\description{Emits ::row_changed for each row in the child model, which causes
the filter to re-evaluate whether a row is visible or not.}
\usage{gtkTreeModelFilterRefilter(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeModelFilter}}]  A \verb{\link{GtkTreeModelFilter}}.}}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
