\alias{gtkTreeModelGetColumnType}
\name{gtkTreeModelGetColumnType}
\title{gtkTreeModelGetColumnType}
\description{Returns the type of the column.}
\usage{gtkTreeModelGetColumnType(object, index)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeModel}}]  A \verb{\link{GtkTreeModel}}.}
\item{\verb{index}}{[integer]  The column index.}
}
\value{[\verb{\link{GType}}]  The type of the column.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
