\alias{gtkTreeModelGetFlags}
\name{gtkTreeModelGetFlags}
\title{gtkTreeModelGetFlags}
\description{Returns a set of flags supported by this interface.  The flags are a bitwise
combination of \code{\link{GtkTreeModelFlags}}.  The flags supported should not change
during the lifecycle of the tree_model.}
\usage{gtkTreeModelGetFlags(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeModel}}]  A \code{\link{GtkTreeModel}}.}}
\value{[\code{\link{GtkTreeModelFlags}}]  The flags supported by this interface.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
