\alias{gtkTreeModelGetPath}
\name{gtkTreeModelGetPath}
\title{gtkTreeModelGetPath}
\description{Returns a newly-created \code{\link{GtkTreePath}} referenced by \code{iter}.}
\usage{gtkTreeModelGetPath(object, iter)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeModel}}]  A \code{\link{GtkTreeModel}}.}
\item{\code{iter}}{[\code{\link{GtkTreeIter}}]  The \code{\link{GtkTreeIter}}.}
}
\value{[\code{\link{GtkTreePath}}]  a newly-created \code{\link{GtkTreePath}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
