\alias{gtkTreeModelIterHasChild}
\name{gtkTreeModelIterHasChild}
\title{gtkTreeModelIterHasChild}
\description{Returns \code{TRUE} if \code{iter} has children, \code{FALSE} otherwise.}
\usage{gtkTreeModelIterHasChild(object, iter)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeModel}}]  A \verb{\link{GtkTreeModel}}.}
\item{\verb{iter}}{[\verb{\link{GtkTreeIter}}]  The \verb{\link{GtkTreeIter}} to test for children.}
}
\value{[logical] \code{TRUE} if \code{iter} has children.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
