\alias{gtkTreeModelRowChanged}
\name{gtkTreeModelRowChanged}
\title{gtkTreeModelRowChanged}
\description{Emits the "row\_changed" signal on \code{tree.model}.}
\usage{gtkTreeModelRowChanged(object, path, iter)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeModel}}]  A \code{\link{GtkTreeModel}}}
\item{\code{path}}{[\code{\link{GtkTreePath}}]  A \code{\link{GtkTreePath}} pointing to the changed row}
\item{\code{iter}}{[\code{\link{GtkTreeIter}}]  A valid \code{\link{GtkTreeIter}} pointing to the changed row}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
