\alias{gtkTreeModelRowInserted}
\name{gtkTreeModelRowInserted}
\title{gtkTreeModelRowInserted}
\description{Emits the "row_inserted" signal on \code{tree.model}}
\usage{gtkTreeModelRowInserted(object, path, iter)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeModel}}]  A \verb{\link{GtkTreeModel}}}
\item{\verb{path}}{[\verb{\link{GtkTreePath}}]  A \verb{\link{GtkTreePath}} pointing to the inserted row}
\item{\verb{iter}}{[\verb{\link{GtkTreeIter}}]  A valid \verb{\link{GtkTreeIter}} pointing to the inserted row}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
