\alias{gtkTreeModelSortGetModel}
\name{gtkTreeModelSortGetModel}
\title{gtkTreeModelSortGetModel}
\description{Returns the model the \verb{\link{GtkTreeModelSort}} is sorting.}
\usage{gtkTreeModelSortGetModel(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeModelSort}}]  a \verb{\link{GtkTreeModelSort}}}}
\value{[\verb{\link{GtkTreeModel}}]  the "child model" being sorted}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
