\alias{gtkTreeModelSortNewWithModel}
\name{gtkTreeModelSortNewWithModel}
\title{gtkTreeModelSortNewWithModel}
\description{Creates a new \verb{\link{GtkTreeModel}}, with \code{child.model} as the child model.}
\usage{gtkTreeModelSortNewWithModel(child.model = NULL)}
\arguments{\item{\verb{child.model}}{[\verb{\link{GtkTreeModel}}]  A \verb{\link{GtkTreeModel}}}}
\value{[\verb{\link{GtkTreeModel}}]  A new \verb{\link{GtkTreeModel}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
