\alias{gtkTreeModelSortResetDefaultSortFunc}
\name{gtkTreeModelSortResetDefaultSortFunc}
\title{gtkTreeModelSortResetDefaultSortFunc}
\description{This resets the default sort function to be in the 'unsorted' state.  That
is, it is in the same order as the child model. It will re-sort the model
to be in the same order as the child model only if the \code{\link{GtkTreeModelSort}}
is in 'unsorted' state.}
\usage{gtkTreeModelSortResetDefaultSortFunc(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeModelSort}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
