\alias{gtkTreePathAppendIndex}
\name{gtkTreePathAppendIndex}
\title{gtkTreePathAppendIndex}
\description{Appends a new index to a path.  As a result, the depth of the path is
increased.}
\usage{gtkTreePathAppendIndex(object, index)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreePath}}]  A \verb{\link{GtkTreePath}}.}
\item{\verb{index}}{[integer]  The index.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
