\alias{gtkTreePathCopy}
\name{gtkTreePathCopy}
\title{gtkTreePathCopy}
\description{Creates a new \code{\link{GtkTreePath}} as a copy of \code{path}.}
\usage{gtkTreePathCopy(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreePath}}]  A \code{\link{GtkTreePath}}.}}
\value{[\code{\link{GtkTreePath}}]  A new \code{\link{GtkTreePath}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
