\alias{gtkTreePathIsAncestor}
\name{gtkTreePathIsAncestor}
\title{gtkTreePathIsAncestor}
\description{Returns \code{TRUE} if \code{descendant} is a descendant of \code{path}.}
\usage{gtkTreePathIsAncestor(object, descendant)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreePath}}]  a \verb{\link{GtkTreePath}}}
\item{\verb{descendant}}{[\verb{\link{GtkTreePath}}]  another \verb{\link{GtkTreePath}}}
}
\value{[logical] \code{TRUE} if \code{descendant} is contained inside \code{path}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
