\alias{gtkTreePathNew}
\name{gtkTreePathNew}
\title{gtkTreePathNew}
\description{Creates a new \verb{\link{GtkTreePath}}.  This structure refers to a row.}
\usage{gtkTreePathNew()}
\value{[\verb{\link{GtkTreePath}}]  A newly created \verb{\link{GtkTreePath}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
