\alias{gtkTreePathNewFirst}
\name{gtkTreePathNewFirst}
\title{gtkTreePathNewFirst}
\description{Creates a new \verb{\link{GtkTreePath}}.  The string representation of this path is "0"}
\usage{gtkTreePathNewFirst()}
\value{[\verb{\link{GtkTreePath}}]  A new \verb{\link{GtkTreePath}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
