\alias{gtkTreePathNewFromIndices}
\name{gtkTreePathNewFromIndices}
\title{gtkTreePathNewFromIndices}
\description{Creates a new path with \code{first.index} and \code{varargs} as indices.}
\usage{gtkTreePathNewFromIndices(...)}
\arguments{\item{\verb{...}}{ list of integers terminated by -1}}
\details{  Since  2.2}
\value{[\verb{\link{GtkTreePath}}]  A newly created \verb{\link{GtkTreePath}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
