\alias{gtkTreePathNext}
\name{gtkTreePathNext}
\title{gtkTreePathNext}
\description{Moves the \code{path} to point to the next node at the current depth.}
\usage{gtkTreePathNext(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreePath}}]  A \code{\link{GtkTreePath}}.}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
