\alias{gtkTreePathPrev}
\name{gtkTreePathPrev}
\title{gtkTreePathPrev}
\description{Moves the \code{path} to point to the previous node at the current depth, 
if it exists.}
\usage{gtkTreePathPrev(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreePath}}.}}
\value{[logical] \code{TRUE} if \code{path} has a previous node, and the move was made.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
