\alias{gtkTreePathToString}
\name{gtkTreePathToString}
\title{gtkTreePathToString}
\description{Generates a string representation of the path.  This string is a ':'
separated list of numbers.  For example, "4:10:0:3" would be an acceptable return value for this string.}
\usage{gtkTreePathToString(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreePath}}}}
\value{[character] A newly-allocated string.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
