\alias{gtkTreePathUp}
\name{gtkTreePathUp}
\title{gtkTreePathUp}
\description{Moves the \code{path} to point to its parent node, if it has a parent.}
\usage{gtkTreePathUp(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreePath}}.}}
\value{[logical] \code{TRUE} if \code{path} has a parent, and the move was made.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
