\alias{gtkTreeRowReferenceCopy}
\name{gtkTreeRowReferenceCopy}
\title{gtkTreeRowReferenceCopy}
\description{Copies a \code{\link{GtkTreeRowReference}}.}
\usage{gtkTreeRowReferenceCopy(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeRowReference}}]  a \code{\link{GtkTreeRowReference}}}}
\details{  Since  2.2}
\value{[\code{\link{GtkTreeRowReference}}]  a copy of \code{reference}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
