\alias{gtkTreeRowReferenceGetPath}
\name{gtkTreeRowReferenceGetPath}
\title{gtkTreeRowReferenceGetPath}
\description{Returns a path that the row reference currently points to, or \code{NULL} if the
path pointed to is no longer valid.}
\usage{gtkTreeRowReferenceGetPath(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeRowReference}}]  A \verb{\link{GtkTreeRowReference}}}}
\value{[\verb{\link{GtkTreePath}}]  A current path, or \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
