\alias{gtkTreeRowReferenceInserted}
\name{gtkTreeRowReferenceInserted}
\title{gtkTreeRowReferenceInserted}
\description{Lets a set of row reference created by \code{\link{gtkTreeRowReferenceNewProxy}}
know that the model emitted the "row\_inserted" signal.}
\usage{gtkTreeRowReferenceInserted(proxy, path)}
\arguments{
\item{\code{proxy}}{[\code{\link{GObject}}]  A \code{\link{GObject}}}
\item{\code{path}}{[\code{\link{GtkTreePath}}]  The row position that was inserted}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
