\alias{gtkTreeRowReferenceValid}
\name{gtkTreeRowReferenceValid}
\title{gtkTreeRowReferenceValid}
\description{Returns TRUE if the \code{reference} is non-NULL and refers to a current valid
path.}
\usage{gtkTreeRowReferenceValid(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeRowReference}}]  A \code{\link{GtkTreeRowReference}}, or NULL}}
\value{[logical]  TRUE if \code{reference} points to a valid path.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
