\alias{gtkTreeSelectionCountSelectedRows}
\name{gtkTreeSelectionCountSelectedRows}
\title{gtkTreeSelectionCountSelectedRows}
\description{Returns the number of rows that have been selected in \code{tree}.}
\usage{gtkTreeSelectionCountSelectedRows(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeSelection}}]  A \code{\link{GtkTreeSelection}}.}}
\details{  Since  2.2}
\value{[integer]  The number of rows selected.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
