\alias{gtkTreeSelectionGetMode}
\name{gtkTreeSelectionGetMode}
\title{gtkTreeSelectionGetMode}
\description{Gets the selection mode for \code{selection}. See
\code{\link{gtkTreeSelectionSetMode}}.}
\usage{gtkTreeSelectionGetMode(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeSelection}}]  a \verb{\link{GtkTreeSelection}}}}
\value{[\code{\link{GtkSelectionMode}}]  the current selection mode}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
