\alias{gtkTreeSelectionGetSelectFunction}
\name{gtkTreeSelectionGetSelectFunction}
\title{gtkTreeSelectionGetSelectFunction}
\description{Returns the current selection function.}
\usage{gtkTreeSelectionGetSelectFunction(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeSelection}}.}}
\details{Since 2.14}
\value{[\code{\link{GtkTreeSelectionFunc}}] The function.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
