\alias{gtkTreeSelectionGetTreeView}
\name{gtkTreeSelectionGetTreeView}
\title{gtkTreeSelectionGetTreeView}
\description{Returns the tree view associated with \code{selection}.}
\usage{gtkTreeSelectionGetTreeView(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeSelection}}]  A \code{\link{GtkTreeSelection}}}}
\value{[\code{\link{GtkTreeView}}]  A \code{\link{GtkTreeView}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
