\alias{gtkTreeSelectionIterIsSelected}
\name{gtkTreeSelectionIterIsSelected}
\title{gtkTreeSelectionIterIsSelected}
\description{Returns \code{TRUE} if the row at \code{iter} is currently selected.}
\usage{gtkTreeSelectionIterIsSelected(object, iter)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeSelection}}]  A \verb{\link{GtkTreeSelection}}}
\item{\verb{iter}}{[\verb{\link{GtkTreeIter}}]  A valid \verb{\link{GtkTreeIter}}}
}
\value{[logical] \code{TRUE}, if \code{iter} is selected}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
