\alias{gtkTreeSelectionSelectAll}
\name{gtkTreeSelectionSelectAll}
\title{gtkTreeSelectionSelectAll}
\description{Selects all the nodes. \code{selection} must be set to \code{GTK_SELECTION_MULTIPLE}
mode.}
\usage{gtkTreeSelectionSelectAll(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeSelection}}]  A \code{\link{GtkTreeSelection}}.}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
