\alias{gtkTreeSelectionSelectIter}
\name{gtkTreeSelectionSelectIter}
\title{gtkTreeSelectionSelectIter}
\description{Selects the specified iterator.}
\usage{gtkTreeSelectionSelectIter(object, iter)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeSelection}}]  A \code{\link{GtkTreeSelection}}.}
\item{\code{iter}}{[\code{\link{GtkTreeIter}}]  The \code{\link{GtkTreeIter}} to be selected.}
}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
