\alias{gtkTreeSelectionSelectPath}
\name{gtkTreeSelectionSelectPath}
\title{gtkTreeSelectionSelectPath}
\description{Select the row at \code{path}.}
\usage{gtkTreeSelectionSelectPath(object, path)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeSelection}}]  A \code{\link{GtkTreeSelection}}.}
\item{\code{path}}{[\code{\link{GtkTreePath}}]  The \code{\link{GtkTreePath}} to be selected.}
}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
