\alias{gtkTreeSelectionSetMode}
\name{gtkTreeSelectionSetMode}
\title{gtkTreeSelectionSetMode}
\description{Sets the selection mode of the \code{selection}.  If the previous type was
\code{GTK_SELECTION_MULTIPLE}, then the anchor is kept selected, if it was
previously selected.}
\usage{gtkTreeSelectionSetMode(object, type)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeSelection}}]  A \code{\link{GtkTreeSelection}}.}
\item{\code{type}}{[\code{\link{GtkSelectionMode}}]  The selection mode}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
