\alias{gtkTreeSelectionUnselectAll}
\name{gtkTreeSelectionUnselectAll}
\title{gtkTreeSelectionUnselectAll}
\description{Unselects all the nodes.}
\usage{gtkTreeSelectionUnselectAll(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeSelection}}]  A \code{\link{GtkTreeSelection}}.}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
