\alias{gtkTreeSelectionUnselectIter}
\name{gtkTreeSelectionUnselectIter}
\title{gtkTreeSelectionUnselectIter}
\description{Unselects the specified iterator.}
\usage{gtkTreeSelectionUnselectIter(object, iter)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeSelection}}]  A \verb{\link{GtkTreeSelection}}.}
\item{\verb{iter}}{[\verb{\link{GtkTreeIter}}]  The \verb{\link{GtkTreeIter}} to be unselected.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
