\alias{gtkTreeSelectionUnselectPath}
\name{gtkTreeSelectionUnselectPath}
\title{gtkTreeSelectionUnselectPath}
\description{Unselects the row at \code{path}.}
\usage{gtkTreeSelectionUnselectPath(object, path)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeSelection}}]  A \code{\link{GtkTreeSelection}}.}
\item{\code{path}}{[\code{\link{GtkTreePath}}]  The \code{\link{GtkTreePath}} to be unselected.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
