\alias{gtkTreeSelectionUnselectRange}
\name{gtkTreeSelectionUnselectRange}
\title{gtkTreeSelectionUnselectRange}
\description{Unselects a range of nodes, determined by \code{start.path} and \code{end.path}
inclusive.}
\usage{gtkTreeSelectionUnselectRange(object, start.path, end.path)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeSelection}}]  A \code{\link{GtkTreeSelection}}.}
\item{\code{start.path}}{[\code{\link{GtkTreePath}}]  The initial node of the range.}
\item{\code{end.path}}{[\code{\link{GtkTreePath}}]  The initial node of the range.}
}
\details{  Since  2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
