\alias{gtkTreeSortableHasDefaultSortFunc}
\name{gtkTreeSortableHasDefaultSortFunc}
\title{gtkTreeSortableHasDefaultSortFunc}
\description{Returns \code{TRUE} if the model has a default sort function. This is used
primarily by GtkTreeViewColumns in order to determine if a model can 
go back to the default state, or not.}
\usage{gtkTreeSortableHasDefaultSortFunc(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeSortable}}}}
\value{[logical] \code{TRUE}, if the model has a default sort function}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
