\alias{gtkTreeSortableSortColumnChanged}
\name{gtkTreeSortableSortColumnChanged}
\title{gtkTreeSortableSortColumnChanged}
\description{Emits a GtkTreeSortable::sort_column_changed signal on}
\usage{gtkTreeSortableSortColumnChanged(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeSortable}}]  A \code{\link{GtkTreeSortable}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
