\alias{gtkTreeStoreClear}
\name{gtkTreeStoreClear}
\title{gtkTreeStoreClear}
\description{Removes all rows from \code{tree.store}}
\usage{gtkTreeStoreClear(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeStore}}]  a \code{\link{GtkTreeStore}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
