\alias{gtkTreeStoreNewv}
\name{gtkTreeStoreNewv}
\title{gtkTreeStoreNewv}
\description{Non vararg creation function.  Used primarily by language bindings.}
\usage{gtkTreeStoreNewv(types)}
\arguments{\item{\code{types}}{[\code{\link{GType}}]  a list of \code{\link{GType}} types for the columns, from first to last}}
\value{[\code{\link{GtkTreeStore}}]  a new \code{\link{GtkTreeStore}}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
