\alias{gtkTreeViewCollapseAll}
\name{gtkTreeViewCollapseAll}
\title{gtkTreeViewCollapseAll}
\description{Recursively collapses all visible, expanded nodes in \code{tree.view}.}
\usage{gtkTreeViewCollapseAll(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeView}}]  A \code{\link{GtkTreeView}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
