\alias{gtkTreeViewColumnClearAttributes}
\name{gtkTreeViewColumnClearAttributes}
\title{gtkTreeViewColumnClearAttributes}
\description{Clears all existing attributes previously set with
\code{\link{gtkTreeViewColumnSetAttributes}}.}
\usage{gtkTreeViewColumnClearAttributes(object, cell.renderer)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeViewColumn}}]  a \verb{\link{GtkTreeViewColumn}}}
\item{\verb{cell.renderer}}{[\verb{\link{GtkCellRenderer}}]  a \verb{\link{GtkCellRenderer}} to clear the attribute mapping on.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
