\alias{gtkTreeViewColumnClicked}
\name{gtkTreeViewColumnClicked}
\title{gtkTreeViewColumnClicked}
\description{Emits the "clicked" signal on the column.  This function will only work if
\code{tree.column} is clickable.}
\usage{gtkTreeViewColumnClicked(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTreeViewColumn}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
