\alias{gtkTreeViewColumnGetAlignment}
\name{gtkTreeViewColumnGetAlignment}
\title{gtkTreeViewColumnGetAlignment}
\description{Returns the current x alignment of \code{tree.column}.  This value can range
between 0.0 and 1.0.}
\usage{gtkTreeViewColumnGetAlignment(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeViewColumn}}]  A \code{\link{GtkTreeViewColumn}}.}}
\value{[numeric]  The current alignent of \code{tree.column}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
