\alias{gtkTreeViewColumnGetClickable}
\name{gtkTreeViewColumnGetClickable}
\title{gtkTreeViewColumnGetClickable}
\description{Returns \code{TRUE} if the user can click on the header for the column.}
\usage{gtkTreeViewColumnGetClickable(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeViewColumn}}]  a \verb{\link{GtkTreeViewColumn}}}}
\value{[logical] \code{TRUE} if user can click the column header.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
