\alias{gtkTreeViewColumnGetFixedWidth}
\name{gtkTreeViewColumnGetFixedWidth}
\title{gtkTreeViewColumnGetFixedWidth}
\description{Gets the fixed width of the column.  This value is only meaning may not be
the actual width of the column on the screen, just what is requested.}
\usage{gtkTreeViewColumnGetFixedWidth(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTreeViewColumn}}}}
\value{[integer] the fixed width of the column}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
