\alias{gtkTreeViewColumnGetMinWidth}
\name{gtkTreeViewColumnGetMinWidth}
\title{gtkTreeViewColumnGetMinWidth}
\description{Returns the minimum width in pixels of the \code{tree.column}, or -1 if no minimum
width is set.}
\usage{gtkTreeViewColumnGetMinWidth(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeViewColumn}}]  A \verb{\link{GtkTreeViewColumn}}.}}
\value{[integer]  The minimum width of the \code{tree.column}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
