\alias{gtkTreeViewColumnGetReorderable}
\name{gtkTreeViewColumnGetReorderable}
\title{gtkTreeViewColumnGetReorderable}
\description{Returns \code{TRUE} if the \code{tree.column} can be reordered by the user.}
\usage{gtkTreeViewColumnGetReorderable(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeViewColumn}}]  A \code{\link{GtkTreeViewColumn}}}}
\value{[logical]  \code{TRUE} if the \code{tree.column} can be reordered by the user.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
