\alias{gtkTreeViewColumnGetResizable}
\name{gtkTreeViewColumnGetResizable}
\title{gtkTreeViewColumnGetResizable}
\description{Returns \code{TRUE} if the \code{tree.column} can be resized by the end user.}
\usage{gtkTreeViewColumnGetResizable(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeViewColumn}}}}
\value{[logical] \code{TRUE}, if the \code{tree.column} can be resized.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
