\alias{gtkTreeViewColumnGetSortIndicator}
\name{gtkTreeViewColumnGetSortIndicator}
\title{gtkTreeViewColumnGetSortIndicator}
\description{Gets the value set by \code{\link{gtkTreeViewColumnSetSortIndicator}}.}
\usage{gtkTreeViewColumnGetSortIndicator(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeViewColumn}}]  a \code{\link{GtkTreeViewColumn}}}}
\value{[logical]  whether the sort indicator arrow is displayed}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
