\alias{gtkTreeViewColumnGetSortOrder}
\name{gtkTreeViewColumnGetSortOrder}
\title{gtkTreeViewColumnGetSortOrder}
\description{Gets the value set by \code{\link{gtkTreeViewColumnSetSortOrder}}.}
\usage{gtkTreeViewColumnGetSortOrder(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeViewColumn}}]  a \code{\link{GtkTreeViewColumn}}}}
\value{[\code{\link{GtkSortType}}]  the sort order the sort indicator is indicating}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
