\alias{gtkTreeViewColumnGetSpacing}
\name{gtkTreeViewColumnGetSpacing}
\title{gtkTreeViewColumnGetSpacing}
\description{Returns the spacing of \code{tree.column}.}
\usage{gtkTreeViewColumnGetSpacing(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeViewColumn}}]  A \code{\link{GtkTreeViewColumn}}.}}
\value{[integer]  the spacing of \code{tree.column}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
