\alias{gtkTreeViewColumnGetTitle}
\name{gtkTreeViewColumnGetTitle}
\title{gtkTreeViewColumnGetTitle}
\description{Returns the title of the widget.}
\usage{gtkTreeViewColumnGetTitle(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeViewColumn}}]  A \code{\link{GtkTreeViewColumn}}.}}
\value{[character]  the title of the column.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
