\alias{gtkTreeViewColumnGetVisible}
\name{gtkTreeViewColumnGetVisible}
\title{gtkTreeViewColumnGetVisible}
\description{Returns \code{TRUE} if \code{tree.column} is visible.}
\usage{gtkTreeViewColumnGetVisible(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeViewColumn}}.}}
\value{[logical] whether the column is visible or not.  If it is visible, then
the tree will show the column.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
