\alias{gtkTreeViewColumnGetWidget}
\name{gtkTreeViewColumnGetWidget}
\title{gtkTreeViewColumnGetWidget}
\description{Returns the \verb{\link{GtkWidget}} in the button on the column header.  If a custom
widget has not been set then \code{NULL} is returned.}
\usage{gtkTreeViewColumnGetWidget(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeViewColumn}}]  A \verb{\link{GtkTreeViewColumn}}.}}
\value{[\verb{\link{GtkWidget}}]  The \verb{\link{GtkWidget}} in the column header, or \code{NULL}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
