\alias{gtkTreeViewColumnGetWidth}
\name{gtkTreeViewColumnGetWidth}
\title{gtkTreeViewColumnGetWidth}
\description{Returns the current size of \code{tree.column} in pixels.}
\usage{gtkTreeViewColumnGetWidth(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeViewColumn}}]  A \code{\link{GtkTreeViewColumn}}.}}
\value{[integer]  The current width of \code{tree.column}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
