\alias{gtkTreeViewColumnNew}
\name{gtkTreeViewColumnNew}
\title{gtkTreeViewColumnNew}
\description{Creates a new \code{\link{GtkTreeViewColumn}}.}
\usage{gtkTreeViewColumnNew()}
\value{[\code{\link{GtkTreeViewColumn}}]  A newly created \code{\link{GtkTreeViewColumn}}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
