\alias{gtkTreeViewColumnQueueResize}
\name{gtkTreeViewColumnQueueResize}
\title{gtkTreeViewColumnQueueResize}
\description{Flags the column, and the cell renderers added to this column, to have
their sizes renegotiated.}
\usage{gtkTreeViewColumnQueueResize(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeViewColumn}}}}
\details{Since 2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
