\alias{gtkTreeViewColumnsAutosize}
\name{gtkTreeViewColumnsAutosize}
\title{gtkTreeViewColumnsAutosize}
\description{Resizes all columns to their optimal width. Only works after the
treeview has been realized.}
\usage{gtkTreeViewColumnsAutosize(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeView}}]  A \code{\link{GtkTreeView}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
