\alias{gtkTreeViewColumnSetAlignment}
\name{gtkTreeViewColumnSetAlignment}
\title{gtkTreeViewColumnSetAlignment}
\description{Sets the alignment of the title or custom widget inside the column header.
The alignment determines its location inside the button -- 0.0 for left, 0.5
for center, 1.0 for right.}
\usage{gtkTreeViewColumnSetAlignment(object, xalign)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeViewColumn}}.}
\item{\verb{xalign}}{The alignment, which is between [0.0 and 1.0] inclusive.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
