\alias{gtkTreeViewColumnSetClickable}
\name{gtkTreeViewColumnSetClickable}
\title{gtkTreeViewColumnSetClickable}
\description{Sets the header to be active if \code{active} is \code{TRUE}.  When the header is active,
then it can take keyboard focus, and can be clicked.}
\usage{gtkTreeViewColumnSetClickable(object, active)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeViewColumn}}]  A \verb{\link{GtkTreeViewColumn}}.}
\item{\verb{active}}{[logical] \code{TRUE} if the header is active.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
