\alias{gtkTreeViewColumnSetMinWidth}
\name{gtkTreeViewColumnSetMinWidth}
\title{gtkTreeViewColumnSetMinWidth}
\description{Sets the minimum width of the \code{tree.column}.  If \code{min.width} is -1, then the
minimum width is unset.}
\usage{gtkTreeViewColumnSetMinWidth(object, min.width)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeViewColumn}}]  A \code{\link{GtkTreeViewColumn}}.}
\item{\code{min.width}}{[integer]  The minimum width of the column in pixels, or -1.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
