\alias{gtkTreeViewColumnSetReorderable}
\name{gtkTreeViewColumnSetReorderable}
\title{gtkTreeViewColumnSetReorderable}
\description{If \code{reorderable} is \code{TRUE}, then the column can be reordered by the end user
dragging the header.}
\usage{gtkTreeViewColumnSetReorderable(object, reorderable)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeViewColumn}}}
\item{\verb{reorderable}}{\code{TRUE}, if the column can be reordered.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
