\alias{gtkTreeViewColumnSetSizing}
\name{gtkTreeViewColumnSetSizing}
\title{gtkTreeViewColumnSetSizing}
\description{Sets the growth behavior of \code{tree.column} to \code{type}.}
\usage{gtkTreeViewColumnSetSizing(object, type)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeViewColumn}}]  A \verb{\link{GtkTreeViewColumn}}.}
\item{\verb{type}}{[\code{\link{GtkTreeViewColumnSizing}}]  The \code{\link{GtkTreeViewColumnSizing}}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
