\alias{gtkTreeViewColumnSetSortColumnId}
\name{gtkTreeViewColumnSetSortColumnId}
\title{gtkTreeViewColumnSetSortColumnId}
\description{Sets the logical \code{sort.column.id} that this column sorts on when this column 
is selected for sorting.  Doing so makes the column header clickable.}
\usage{gtkTreeViewColumnSetSortColumnId(object, sort.column.id)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTreeViewColumn}}}
\item{\verb{sort.column.id}}{The \code{sort.column.id} of the model to sort on.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
