\alias{gtkTreeViewColumnSetSortIndicator}
\name{gtkTreeViewColumnSetSortIndicator}
\title{gtkTreeViewColumnSetSortIndicator}
\description{Call this function with a \code{setting} of \code{TRUE} to display an arrow in
the header button indicating the column is sorted. Call
\code{\link{gtkTreeViewColumnSetSortOrder}} to change the direction of
the arrow.}
\usage{gtkTreeViewColumnSetSortIndicator(object, setting)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTreeViewColumn}}}
\item{\verb{setting}}{\code{TRUE} to display an indicator that the column is sorted}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
