\alias{gtkTreeViewColumnSetSpacing}
\name{gtkTreeViewColumnSetSpacing}
\title{gtkTreeViewColumnSetSpacing}
\description{Sets the spacing field of \code{tree.column}, which is the number of pixels to
place between cell renderers packed into it.}
\usage{gtkTreeViewColumnSetSpacing(object, spacing)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeViewColumn}}]  A \code{\link{GtkTreeViewColumn}}.}
\item{\code{spacing}}{[integer]  distance between cell renderers in pixels.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
