\alias{gtkTreeViewColumnSetVisible}
\name{gtkTreeViewColumnSetVisible}
\title{gtkTreeViewColumnSetVisible}
\description{Sets the visibility of \code{tree.column}.}
\usage{gtkTreeViewColumnSetVisible(object, visible)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeViewColumn}}.}
\item{\verb{visible}}{\code{TRUE} if the \code{tree.column} is visible.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
