\alias{gtkTreeViewColumnSetWidget}
\name{gtkTreeViewColumnSetWidget}
\title{gtkTreeViewColumnSetWidget}
\description{Sets the widget in the header to be \code{widget}.  If widget is \code{NULL}, then the
header button is set with a \verb{\link{GtkLabel}} set to the title of \code{tree.column}.}
\usage{gtkTreeViewColumnSetWidget(object, widget = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeViewColumn}}]  A \verb{\link{GtkTreeViewColumn}}.}
\item{\verb{widget}}{[\verb{\link{GtkWidget}}]  A child \verb{\link{GtkWidget}}, or \code{NULL}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
