\alias{gtkTreeViewExpandToPath}
\name{gtkTreeViewExpandToPath}
\title{gtkTreeViewExpandToPath}
\description{Expands the row at \code{path}. This will also expand all parent rows of
\code{path} as necessary.}
\usage{gtkTreeViewExpandToPath(object, path)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeView}}]  A \code{\link{GtkTreeView}}.}
\item{\code{path}}{[\code{\link{GtkTreePath}}]  path to a row.}
}
\details{  Since  2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
