\alias{gtkTreeViewGetColumns}
\name{gtkTreeViewGetColumns}
\title{gtkTreeViewGetColumns}
\description{Returns a \verb{list} of all the \verb{\link{GtkTreeViewColumn}} s currently in \code{tree.view}.}
\usage{gtkTreeViewGetColumns(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeView}}]  A \verb{\link{GtkTreeView}}}}
\value{[list]  A list of \verb{\link{GtkTreeViewColumn}} s}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
