\alias{gtkTreeViewGetEnableSearch}
\name{gtkTreeViewGetEnableSearch}
\title{gtkTreeViewGetEnableSearch}
\description{Returns whether or not the tree allows to start interactive searching 
by typing in text.}
\usage{gtkTreeViewGetEnableSearch(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeView}}}}
\value{[logical] whether or not to let the user search interactively}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
