\alias{gtkTreeViewGetEnableTreeLines}
\name{gtkTreeViewGetEnableTreeLines}
\title{gtkTreeViewGetEnableTreeLines}
\description{Returns whether or not tree lines are drawn in \code{tree.view}.}
\usage{gtkTreeViewGetEnableTreeLines(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeView}}]  a \verb{\link{GtkTreeView}}.}}
\details{  Since  2.10}
\value{[logical] \code{TRUE} if tree lines are drawn in \code{tree.view}, \code{FALSE}
otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
