\alias{gtkTreeViewGetExpanderColumn}
\name{gtkTreeViewGetExpanderColumn}
\title{gtkTreeViewGetExpanderColumn}
\description{Returns the column that is the current expander column.  This
column has the expander arrow drawn next to it.}
\usage{gtkTreeViewGetExpanderColumn(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeView}}}}
\value{[\code{\link{GtkTreeViewColumn}}] The expander column.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
