\alias{gtkTreeViewGetFixedHeightMode}
\name{gtkTreeViewGetFixedHeightMode}
\title{gtkTreeViewGetFixedHeightMode}
\description{Returns whether fixed height mode is turned on for \code{tree.view}.}
\usage{gtkTreeViewGetFixedHeightMode(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTreeView}}}}
\details{Since 2.6}
\value{[logical] \code{TRUE} if \code{tree.view} is in fixed height mode}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
