\alias{gtkTreeViewGetGridLines}
\name{gtkTreeViewGetGridLines}
\title{gtkTreeViewGetGridLines}
\description{Returns which grid lines are enabled in \code{tree.view}.}
\usage{gtkTreeViewGetGridLines(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTreeView}}}}
\details{Since 2.10}
\value{[\code{\link{GtkTreeViewGridLines}}] a \code{\link{GtkTreeViewGridLines}} value indicating which grid lines
are enabled.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
