\alias{gtkTreeViewGetHadjustment}
\name{gtkTreeViewGetHadjustment}
\title{gtkTreeViewGetHadjustment}
\description{Gets the \verb{\link{GtkAdjustment}} currently being used for the horizontal aspect.}
\usage{gtkTreeViewGetHadjustment(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeView}}]  A \verb{\link{GtkTreeView}}}}
\value{[\verb{\link{GtkAdjustment}}]  A \verb{\link{GtkAdjustment}} object, or \code{NULL} if none is currently being
used.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
