\alias{gtkTreeViewGetHeadersClickable}
\name{gtkTreeViewGetHeadersClickable}
\title{gtkTreeViewGetHeadersClickable}
\description{Returns whether all header columns are clickable.}
\usage{gtkTreeViewGetHeadersClickable(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeView}}]  A \verb{\link{GtkTreeView}}.}}
\details{  Since  2.10}
\value{[logical] \code{TRUE} if all header columns are clickable, otherwise \code{FALSE}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
