\alias{gtkTreeViewGetHeadersVisible}
\name{gtkTreeViewGetHeadersVisible}
\title{gtkTreeViewGetHeadersVisible}
\description{Returns \code{TRUE} if the headers on the \code{tree.view} are visible.}
\usage{gtkTreeViewGetHeadersVisible(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeView}}.}}
\value{[logical] Whether the headers are visible or not.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
