\alias{gtkTreeViewGetHoverExpand}
\name{gtkTreeViewGetHoverExpand}
\title{gtkTreeViewGetHoverExpand}
\description{Returns whether hover expansion mode is turned on for \code{tree.view}.}
\usage{gtkTreeViewGetHoverExpand(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTreeView}}}}
\details{Since 2.6}
\value{[logical] \code{TRUE} if \code{tree.view} is in hover expansion mode}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
