\alias{gtkTreeViewGetHoverSelection}
\name{gtkTreeViewGetHoverSelection}
\title{gtkTreeViewGetHoverSelection}
\description{Returns whether hover selection mode is turned on for \code{tree.view}.}
\usage{gtkTreeViewGetHoverSelection(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTreeView}}}}
\details{Since 2.6}
\value{[logical] \code{TRUE} if \code{tree.view} is in hover selection mode}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
