\alias{gtkTreeViewGetLevelIndentation}
\name{gtkTreeViewGetLevelIndentation}
\title{gtkTreeViewGetLevelIndentation}
\description{Returns the amount, in pixels, of extra indentation for child levels
in \code{tree.view}.}
\usage{gtkTreeViewGetLevelIndentation(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTreeView}}.}}
\details{Since 2.12}
\value{[integer] the amount of extra indentation for child levels in
\code{tree.view}.  A return value of 0 means that this feature is disabled.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
