\alias{gtkTreeViewGetModel}
\name{gtkTreeViewGetModel}
\title{gtkTreeViewGetModel}
\description{Returns the model the \verb{\link{GtkTreeView}} is based on.  Returns \code{NULL} if the
model is unset.}
\usage{gtkTreeViewGetModel(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeView}}]  a \verb{\link{GtkTreeView}}}}
\value{[\verb{\link{GtkTreeModel}}]  A \verb{\link{GtkTreeModel}}, or \code{NULL} if none is currently being used.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
