\alias{gtkTreeViewGetReorderable}
\name{gtkTreeViewGetReorderable}
\title{gtkTreeViewGetReorderable}
\description{Retrieves whether the user can reorder the tree via drag-and-drop. See
\code{\link{gtkTreeViewSetReorderable}}.}
\usage{gtkTreeViewGetReorderable(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTreeView}}}}
\value{[logical] \code{TRUE} if the tree can be reordered.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
