\alias{gtkTreeViewGetRowSeparatorFunc}
\name{gtkTreeViewGetRowSeparatorFunc}
\title{gtkTreeViewGetRowSeparatorFunc}
\description{Returns the current row separator function.}
\usage{gtkTreeViewGetRowSeparatorFunc(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeView}}]  a \code{\link{GtkTreeView}}}}
\details{  Since  2.6}
\value{[\code{\link{GtkTreeViewRowSeparatorFunc}}]  the current row separator function.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
