\alias{gtkTreeViewGetRubberBanding}
\name{gtkTreeViewGetRubberBanding}
\title{gtkTreeViewGetRubberBanding}
\description{Returns whether rubber banding is turned on for \code{tree.view}.  If the
selection mode is \verb{GTK_SELECTION_MULTIPLE}, rubber banding will allow the
user to select multiple rows by dragging the mouse.}
\usage{gtkTreeViewGetRubberBanding(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTreeView}}}}
\details{Since 2.10}
\value{[logical] \code{TRUE} if rubber banding in \code{tree.view} is enabled.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
