\alias{gtkTreeViewGetRulesHint}
\name{gtkTreeViewGetRulesHint}
\title{gtkTreeViewGetRulesHint}
\description{Gets the setting set by \code{\link{gtkTreeViewSetRulesHint}}.}
\usage{gtkTreeViewGetRulesHint(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeView}}]  a \code{\link{GtkTreeView}}}}
\value{[logical]  \code{TRUE} if rules are useful for the user of this tree}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
