\alias{gtkTreeViewGetSelection}
\name{gtkTreeViewGetSelection}
\title{gtkTreeViewGetSelection}
\description{Gets the \verb{\link{GtkTreeSelection}} associated with \code{tree.view}.}
\usage{gtkTreeViewGetSelection(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeView}}]  A \verb{\link{GtkTreeView}}.}}
\value{[\verb{\link{GtkTreeSelection}}]  A \verb{\link{GtkTreeSelection}} object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
