\alias{gtkTreeViewGetTooltipColumn}
\name{gtkTreeViewGetTooltipColumn}
\title{gtkTreeViewGetTooltipColumn}
\description{Returns the column of \code{tree.view}'s model which is being used for
displaying tooltips on \code{tree.view}'s rows.}
\usage{gtkTreeViewGetTooltipColumn(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTreeView}}}}
\details{Since 2.12}
\value{[integer] the index of the tooltip column that is currently being
used, or -1 if this is disabled.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
