\alias{gtkTreeViewGetVadjustment}
\name{gtkTreeViewGetVadjustment}
\title{gtkTreeViewGetVadjustment}
\description{Gets the \verb{\link{GtkAdjustment}} currently being used for the vertical aspect.}
\usage{gtkTreeViewGetVadjustment(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeView}}]  A \verb{\link{GtkTreeView}}}}
\value{[\verb{\link{GtkAdjustment}}]  A \verb{\link{GtkAdjustment}} object, or \code{NULL} if none is currently being
used.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
