\alias{gtkTreeViewIsRubberBandingActive}
\name{gtkTreeViewIsRubberBandingActive}
\title{gtkTreeViewIsRubberBandingActive}
\description{Returns whether a rubber banding operation is currently being done
in \code{tree.view}.}
\usage{gtkTreeViewIsRubberBandingActive(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTreeView}}}}
\details{Since 2.12}
\value{[logical] \code{TRUE} if a rubber banding operation is currently being
done in \code{tree.view}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
