\alias{gtkTreeViewMapExpandedRows}
\name{gtkTreeViewMapExpandedRows}
\title{gtkTreeViewMapExpandedRows}
\description{Calls \code{func} on all expanded rows.}
\usage{gtkTreeViewMapExpandedRows(object, func, data = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeView}}]  A \code{\link{GtkTreeView}}}
\item{\code{func}}{[\code{\link{GtkTreeViewMappingFunc}}]  A function to be called}
\item{\code{data}}{[R object]  User data to be passed to the function.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
