\alias{gtkTreeViewNewWithModel}
\name{gtkTreeViewNewWithModel}
\title{gtkTreeViewNewWithModel}
\description{Creates a new \code{\link{GtkTreeView}} widget with the model initialized to \code{model}.}
\usage{gtkTreeViewNewWithModel(model = NULL, show = TRUE)}
\arguments{\item{\code{model}}{[\code{\link{GtkTreeModel}}]  the model.}}
\value{[\code{\link{GtkWidget}}]  A newly created \code{\link{GtkTreeView}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
