\alias{gtkTreeViewRowActivated}
\name{gtkTreeViewRowActivated}
\title{gtkTreeViewRowActivated}
\description{Activates the cell determined by \code{path} and \code{column}.}
\usage{gtkTreeViewRowActivated(object, path, column)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeView}}]  A \code{\link{GtkTreeView}}}
\item{\code{path}}{[\code{\link{GtkTreePath}}]  The \code{\link{GtkTreePath}} to be activated.}
\item{\code{column}}{[\code{\link{GtkTreeViewColumn}}]  The \code{\link{GtkTreeViewColumn}} to be activated.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
