\alias{gtkTreeViewSetDragDestRow}
\name{gtkTreeViewSetDragDestRow}
\title{gtkTreeViewSetDragDestRow}
\description{Sets the row that is highlighted for feedback.}
\usage{gtkTreeViewSetDragDestRow(object, path, pos)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeView}}]  a \code{\link{GtkTreeView}}}
\item{\code{path}}{[\code{\link{GtkTreePath}}]  The path of the row to highlight, or \code{NULL}.}
\item{\code{pos}}{[\code{\link{GtkTreeViewDropPosition}}]  Specifies whether to drop before, after or into the row}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
