\alias{gtkTreeViewSetEnableTreeLines}
\name{gtkTreeViewSetEnableTreeLines}
\title{gtkTreeViewSetEnableTreeLines}
\description{Sets whether to draw lines interconnecting the expanders in \code{tree.view}.
This does not have any visible effects for lists.}
\usage{gtkTreeViewSetEnableTreeLines(object, enabled)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTreeView}}}
\item{\verb{enabled}}{\code{TRUE} to enable tree line drawing, \code{FALSE} otherwise.}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
