\alias{gtkTreeViewSetHadjustment}
\name{gtkTreeViewSetHadjustment}
\title{gtkTreeViewSetHadjustment}
\description{Sets the \verb{\link{GtkAdjustment}} for the current horizontal aspect.}
\usage{gtkTreeViewSetHadjustment(object, adjustment)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeView}}]  A \verb{\link{GtkTreeView}}}
\item{\verb{adjustment}}{[\verb{\link{GtkAdjustment}}]  The \verb{\link{GtkAdjustment}} to set, or \code{NULL}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
