\alias{gtkTreeViewSetHeadersClickable}
\name{gtkTreeViewSetHeadersClickable}
\title{gtkTreeViewSetHeadersClickable}
\description{Allow the column title buttons to be clicked.}
\usage{gtkTreeViewSetHeadersClickable(object, active)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeView}}]  A \code{\link{GtkTreeView}}.}
\item{\code{active}}{[logical]  \code{TRUE} if the columns are clickable.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
