\alias{gtkTreeViewSetHeadersVisible}
\name{gtkTreeViewSetHeadersVisible}
\title{gtkTreeViewSetHeadersVisible}
\description{Sets the visibility state of the headers.}
\usage{gtkTreeViewSetHeadersVisible(object, headers.visible)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeView}}]  A \code{\link{GtkTreeView}}.}
\item{\code{headers.visible}}{[logical]  \code{TRUE} if the headers are visible}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
