\alias{gtkTreeViewSetRubberBanding}
\name{gtkTreeViewSetRubberBanding}
\title{gtkTreeViewSetRubberBanding}
\description{Enables or disables rubber banding in \code{tree.view}.  If the selection mode
is \code{GTK_SELECTION_MULTIPLE}, rubber banding will allow the user to select
multiple rows by dragging the mouse.}
\usage{gtkTreeViewSetRubberBanding(object, enable)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeView}}]  a \code{\link{GtkTreeView}}}
\item{\code{enable}}{[logical] \code{TRUE} to enable rubber banding}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
