\alias{gtkTreeViewSetTooltipRow}
\name{gtkTreeViewSetTooltipRow}
\title{gtkTreeViewSetTooltipRow}
\description{Sets the tip area of \code{tooltip} to be the area covered by the row at \code{path}.
See also \code{\link{gtkTooltipsetTipArea}}.}
\usage{gtkTreeViewSetTooltipRow(object, tooltip, path)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeView}}]  a \code{\link{GtkTreeView}}}
\item{\code{tooltip}}{[\code{\link{GtkTooltip}}]  a \code{\link{GtkTooltip}}}
\item{\code{path}}{[\code{\link{GtkTreePath}}]  a \code{\link{GtkTreePath}}}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
