\alias{gtkTreeViewSetVadjustment}
\name{gtkTreeViewSetVadjustment}
\title{gtkTreeViewSetVadjustment}
\description{Sets the \verb{\link{GtkAdjustment}} for the current vertical aspect.}
\usage{gtkTreeViewSetVadjustment(object, adjustment)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeView}}]  A \verb{\link{GtkTreeView}}}
\item{\verb{adjustment}}{[\verb{\link{GtkAdjustment}}]  The \verb{\link{GtkAdjustment}} to set, or \code{NULL}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
