\alias{gtkTreeViewUnsetRowsDragDest}
\name{gtkTreeViewUnsetRowsDragDest}
\title{gtkTreeViewUnsetRowsDragDest}
\description{Undoes the effect of \code{\link{gtkTreeViewEnableModelDragDest}}.}
\usage{gtkTreeViewUnsetRowsDragDest(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeView}}]  a \verb{\link{GtkTreeView}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
