\alias{gtkTreeViewUnsetRowsDragSource}
\name{gtkTreeViewUnsetRowsDragSource}
\title{gtkTreeViewUnsetRowsDragSource}
\description{Undoes the effect of \code{\link{gtkTreeViewEnableModelDragSource}}.}
\usage{gtkTreeViewUnsetRowsDragSource(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeView}}]  a \code{\link{GtkTreeView}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
