\alias{gtkUIManagerGetAccelGroup}
\name{gtkUIManagerGetAccelGroup}
\title{gtkUIManagerGetAccelGroup}
\description{Returns the \verb{\link{GtkAccelGroup}} associated with \code{self}.}
\usage{gtkUIManagerGetAccelGroup(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkUIManager}}]  a \verb{\link{GtkUIManager}} object}}
\details{  Since  2.4}
\value{[\verb{\link{GtkAccelGroup}}]  the \verb{\link{GtkAccelGroup}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
