\alias{gtkUIManagerGetUi}
\name{gtkUIManagerGetUi}
\title{gtkUIManagerGetUi}
\description{Creates a UI definition of the merged UI.}
\usage{gtkUIManagerGetUi(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkUIManager}}}}
\details{Since 2.4}
\value{[character] A newly allocated string containing an XML representation of
the merged UI.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
