\alias{gtkVButtonBoxGetLayoutDefault}
\name{gtkVButtonBoxGetLayoutDefault}
\title{gtkVButtonBoxGetLayoutDefault}
\description{
Retrieves the current layout used to arrange buttons in button box widgets.
\strong{WARNING: \code{gtk_vbutton_box_get_layout_default} is deprecated and should not be used in newly-written code.}
}
\usage{gtkVButtonBoxGetLayoutDefault()}
\value{[\code{\link{GtkButtonBoxStyle}}] the current \code{\link{GtkButtonBoxStyle}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
