\alias{gtkVButtonBoxGetSpacingDefault}
\name{gtkVButtonBoxGetSpacingDefault}
\title{gtkVButtonBoxGetSpacingDefault}
\description{
Retrieves the current default spacing for vertical button boxes. This is the number of pixels
to be placed between the buttons when they are arranged.
\strong{WARNING: \code{gtk_vbutton_box_get_spacing_default} is deprecated and should not be used in newly-written code.}
}
\usage{gtkVButtonBoxGetSpacingDefault()}
\value{[integer] the default number of pixels between buttons.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
