\alias{gtkVButtonBoxSetLayoutDefault}
\name{gtkVButtonBoxSetLayoutDefault}
\title{gtkVButtonBoxSetLayoutDefault}
\description{
Sets a new layout mode that will be used by all button boxes.
\strong{WARNING: \code{gtk_vbutton_box_set_layout_default} is deprecated and should not be used in newly-written code.}
}
\usage{gtkVButtonBoxSetLayoutDefault(layout)}
\arguments{\item{\verb{layout}}{a new \code{\link{GtkButtonBoxStyle}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
