\alias{gtkVButtonBoxSetSpacingDefault}
\name{gtkVButtonBoxSetSpacingDefault}
\title{gtkVButtonBoxSetSpacingDefault}
\description{
Changes the default spacing that is placed between widgets in an
vertical button box.
\strong{WARNING: \code{gtk_vbutton_box_set_spacing_default} is deprecated and should not be used in newly-written code.}
}
\usage{gtkVButtonBoxSetSpacingDefault(spacing)}
\arguments{\item{\verb{spacing}}{an integer value.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
