\alias{gtkViewportGetBinWindow}
\name{gtkViewportGetBinWindow}
\title{gtkViewportGetBinWindow}
\description{Gets the bin window of the \code{\link{GtkViewport}}.}
\usage{gtkViewportGetBinWindow(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkViewport}}}}
\details{Since 2.20}
\value{[\code{\link{GdkWindow}}] a \code{\link{GdkWindow}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
