\alias{gtkViewportGetHadjustment}
\name{gtkViewportGetHadjustment}
\title{gtkViewportGetHadjustment}
\description{Returns the horizontal adjustment of the viewport.}
\usage{gtkViewportGetHadjustment(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkViewport}}]  a \code{\link{GtkViewport}}.}}
\value{[\code{\link{GtkAdjustment}}]  the horizontal adjustment of \code{viewport}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
