\alias{gtkViewportGetShadowType}
\name{gtkViewportGetShadowType}
\title{gtkViewportGetShadowType}
\description{Gets the shadow type of the \verb{\link{GtkViewport}}. See
\code{\link{gtkViewportSetShadowType}}.}
\usage{gtkViewportGetShadowType(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkViewport}}]  a \verb{\link{GtkViewport}}}}
\value{[\code{\link{GtkShadowType}}]  the shadow type }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
