\alias{gtkViewportNew}
\name{gtkViewportNew}
\title{gtkViewportNew}
\description{Creates a new \verb{\link{GtkViewport}} with the given adjustments.}
\usage{gtkViewportNew(hadjustment = NULL, vadjustment = NULL, show = TRUE)}
\arguments{
\item{\verb{hadjustment}}{[\verb{\link{GtkAdjustment}}]  horizontal adjustment.}
\item{\verb{vadjustment}}{[\verb{\link{GtkAdjustment}}]  vertical adjustment.}
}
\value{[\verb{\link{GtkWidget}}]  a new \verb{\link{GtkViewport}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
