\alias{gtkViewportSetShadowType}
\name{gtkViewportSetShadowType}
\title{gtkViewportSetShadowType}
\description{Sets the shadow type of the viewport.}
\usage{gtkViewportSetShadowType(object, type)}
\arguments{
\item{\code{object}}{[\code{\link{GtkViewport}}]  a \code{\link{GtkViewport}}.}
\item{\code{type}}{[\code{\link{GtkShadowType}}]  the new shadow type.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
