\alias{gtkViewportSetVadjustment}
\name{gtkViewportSetVadjustment}
\title{gtkViewportSetVadjustment}
\description{Sets the vertical adjustment of the viewport.}
\usage{gtkViewportSetVadjustment(object, adjustment = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkViewport}}]  a \code{\link{GtkViewport}}.}
\item{\code{adjustment}}{[\code{\link{GtkAdjustment}}]  a \code{\link{GtkAdjustment}}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
