\alias{gtkVRulerNew}
\name{gtkVRulerNew}
\title{gtkVRulerNew}
\description{Creates a new vertical ruler}
\usage{gtkVRulerNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkVRuler}}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
