\alias{gtkVScaleNew}
\name{gtkVScaleNew}
\title{gtkVScaleNew}
\description{Creates a new \code{\link{GtkVScale}}.}
\usage{gtkVScaleNew(adjustment = NULL, show = TRUE)}
\arguments{\item{\code{adjustment}}{[\code{\link{GtkAdjustment}}] the \code{\link{GtkAdjustment}} which sets the range of the scale.}}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkVScale}}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
