\alias{GtkVScrollbar}
\alias{gtkVScrollbar}
\name{GtkVScrollbar}
\title{GtkVScrollbar}
\description{\emph{undocumented
}}
\section{Methods and Functions}{
\code{\link{gtkVScrollbarNew}(adjustment = NULL, show = TRUE)}\cr
\code{gtkVScrollbar(adjustment = NULL, show = TRUE)}
}
\section{Hierarchy}{\preformatted{GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkRange
                           +----GtkScrollbar
                                 +----GtkVScrollbar}}
\section{Interfaces}{GtkVScrollbar implements
 AtkImplementorIface and  \verb{\link{GtkBuildable}}.}
\section{Structures}{\describe{\item{\verb{GtkVScrollbar}}{
\emph{undocumented
}

}}}
\section{Convenient Construction}{\code{gtkVScrollbar} is the equivalent of \code{\link{gtkVScrollbarNew}}.}
\references{\url{http://developer.gnome.org/doc/API/2.0/gtk/GtkVScrollbar.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
