\alias{gtkVSeparatorNew}
\name{gtkVSeparatorNew}
\title{gtkVSeparatorNew}
\description{Creates a new \verb{\link{GtkVSeparator}}.}
\usage{gtkVSeparatorNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}] a new \verb{\link{GtkVSeparator}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
