\alias{gtkWidgetClassInstallStylePropertyParser}
\name{gtkWidgetClassInstallStylePropertyParser}
\title{gtkWidgetClassInstallStylePropertyParser}
\description{Installs a style property on a widget class.}
\usage{gtkWidgetClassInstallStylePropertyParser(klass, pspec, parser)}
\arguments{
\item{\code{klass}}{[\code{\link{GtkWidgetClass}}]  a \code{\link{GtkWidgetClass}}}
\item{\code{pspec}}{[\code{\link{GParamSpec}}]  the \code{\link{GParamSpec}} for the style property}
\item{\code{parser}}{[GtkRcPropertyParser]  the parser for the style property}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
