\alias{gtkWidgetCreatePangoContext}
\name{gtkWidgetCreatePangoContext}
\title{gtkWidgetCreatePangoContext}
\description{Creates a new \verb{\link{PangoContext}} with the appropriate font map,
font description, and base direction for drawing text for
this widget. See also \code{\link{gtkWidgetGetPangoContext}}.}
\usage{gtkWidgetCreatePangoContext(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}}}
\value{[\verb{\link{PangoContext}}]  the new \verb{\link{PangoContext}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
