\alias{gtkWidgetEnsureStyle}
\name{gtkWidgetEnsureStyle}
\title{gtkWidgetEnsureStyle}
\description{Ensures that \code{widget} has a style (\code{widget->style}). Not a very useful
function; most of the time, if you want the style, the widget is
realized, and realized widgets are guaranteed to have a style
already.}
\usage{gtkWidgetEnsureStyle(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
