\name{GtkObjectFlags}
\alias{gtkObjectFlags}
\alias{gtkWidgetGetFlags}
\alias{gtkWidgetSetFlags}
\alias{gtkWidgetUnsetFlags}
\title{Set and get GtkWidget flag values.}
\description{
  These functions allow one retrieve GtkObject flags and set them on GtkWidgets.
}
\usage{
gtkObjectFlags(object)
gtkWidgetSetFlags(wid, flags)
gtkWidgetUnsetFlags(wid, flags)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{the object for the flags query}
  \item{wid}{the widget}
  \item{flags}{the flags to (un)set on the widget}
}

\author{Michael Lawrence}
\keyword{interface}
\keyword{internal}


