\alias{gtkWidgetGetAllocation}
\name{gtkWidgetGetAllocation}
\title{gtkWidgetGetAllocation}
\description{Retrieves the widget's allocation.}
\usage{gtkWidgetGetAllocation(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.18}
\value{
A list containing the following elements:
\item{\verb{allocation}}{a pointer to a \code{\link{GtkAllocation}} to copy to. \emph{[  \acronym{out}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
