\alias{gtkWidgetGetAppPaintable}
\name{gtkWidgetGetAppPaintable}
\title{gtkWidgetGetAppPaintable}
\description{Determines whether the application intends to draw on the widget in
an \verb{"expose-event"} handler.}
\usage{gtkWidgetGetAppPaintable(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{See \code{\link{gtkWidgetSetAppPaintable}}
  Since 2.18}
\value{[logical] \code{TRUE} if the widget is app paintable}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
