\alias{gtkWidgetGetCanDefault}
\name{gtkWidgetGetCanDefault}
\title{gtkWidgetGetCanDefault}
\description{Determines whether \code{widget} can be a default widget. See
\code{\link{gtkWidgetSetCanDefault}}.}
\usage{gtkWidgetGetCanDefault(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.18}
\value{[logical] \code{TRUE} if \code{widget} can be a default widget, \code{FALSE} otherwise}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
