\alias{gtkWidgetGetCanFocus}
\name{gtkWidgetGetCanFocus}
\title{gtkWidgetGetCanFocus}
\description{Determines whether \code{widget} can own the input focus. See
\code{\link{gtkWidgetSetCanFocus}}.}
\usage{gtkWidgetGetCanFocus(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.18}
\value{[logical] \code{TRUE} if \code{widget} can own the input focus, \code{FALSE} otherwise}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
