\alias{gtkWidgetGetColormap}
\name{gtkWidgetGetColormap}
\title{gtkWidgetGetColormap}
\description{Gets the colormap that will be used to render \code{widget}. No reference will
be added to the returned colormap; it should not be unreferenced.}
\usage{gtkWidgetGetColormap(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}}}
\value{[\verb{\link{GdkColormap}}]  the colormap used by \code{widget}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
