\alias{gtkWidgetGetCompositeName}
\name{gtkWidgetGetCompositeName}
\title{gtkWidgetGetCompositeName}
\description{Obtains the composite name of a widget.}
\usage{gtkWidgetGetCompositeName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\value{[character] the composite name of \code{widget}, or \code{NULL} if \code{widget} is not
a composite child.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
