\alias{gtkWidgetGetDefaultDirection}
\name{gtkWidgetGetDefaultDirection}
\title{gtkWidgetGetDefaultDirection}
\description{Obtains the current default reading direction. See
\code{\link{gtkWidgetSetDefaultDirection}}.}
\usage{gtkWidgetGetDefaultDirection()}
\value{[\code{\link{GtkTextDirection}}]  the current default direction. }
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
