\alias{gtkWidgetGetDefaultVisual}
\name{gtkWidgetGetDefaultVisual}
\title{gtkWidgetGetDefaultVisual}
\description{Obtains the visual of the default colormap. Not really useful;
used to be useful before \code{\link{gdkColormapGetVisual}} existed.}
\usage{gtkWidgetGetDefaultVisual()}
\value{[\verb{\link{GdkVisual}}]  visual of the default colormap}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
