\alias{gtkWidgetGetDirection}
\name{gtkWidgetGetDirection}
\title{gtkWidgetGetDirection}
\description{Gets the reading direction for a particular widget. See
\code{\link{gtkWidgetSetDirection}}.}
\usage{gtkWidgetGetDirection(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\value{[\code{\link{GtkTextDirection}}] the reading direction for the widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
