\alias{gtkWidgetGetDoubleBuffered}
\name{gtkWidgetGetDoubleBuffered}
\title{gtkWidgetGetDoubleBuffered}
\description{Determines whether the widget is double buffered.}
\usage{gtkWidgetGetDoubleBuffered(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{See \code{\link{gtkWidgetSetDoubleBuffered}}
  Since 2.18}
\value{[logical] \code{TRUE} if the widget is double buffered}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
