\alias{gtkWidgetGetEvents}
\name{gtkWidgetGetEvents}
\title{gtkWidgetGetEvents}
\description{Returns the event mask for the widget (a bitfield containing flags
from the \code{\link{GdkEventMask}} enumeration). These are the events that the widget
will receive.}
\usage{gtkWidgetGetEvents(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\value{[integer] event mask for \code{widget}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
