\alias{gtkWidgetGetExtensionEvents}
\name{gtkWidgetGetExtensionEvents}
\title{gtkWidgetGetExtensionEvents}
\description{Retrieves the extension events the widget will receive; see
\code{\link{gdkInputSetExtensionEvents}}.}
\usage{gtkWidgetGetExtensionEvents(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkWidget}}]  a \code{\link{GtkWidget}}}}
\value{[\code{\link{GdkExtensionMode}}]  extension events for \code{widget}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
