\alias{gtkWidgetGetHasTooltip}
\name{gtkWidgetGetHasTooltip}
\title{gtkWidgetGetHasTooltip}
\description{Returns the current value of the has-tooltip property.  See
GtkWidget:has-tooltip for more information.}
\usage{gtkWidgetGetHasTooltip(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.12}
\value{[logical] current value of has-tooltip on \code{widget}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
