\alias{gtkWidgetGetHasWindow}
\name{gtkWidgetGetHasWindow}
\title{gtkWidgetGetHasWindow}
\description{Determines whether \code{widget} has a \code{\link{GdkWindow}} of its own. See
\code{\link{gtkWidgetSetHasWindow}}.}
\usage{gtkWidgetGetHasWindow(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.18}
\value{[logical] \code{TRUE} if \code{widget} has a window, \code{FALSE} otherwise}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
