\alias{gtkWidgetGetMapped}
\name{gtkWidgetGetMapped}
\title{gtkWidgetGetMapped}
\description{Whether the widget is mapped.}
\usage{gtkWidgetGetMapped(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.20}
\value{[logical] \code{TRUE} if the widget is mapped, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
