\alias{gtkWidgetGetNoShowAll}
\name{gtkWidgetGetNoShowAll}
\title{gtkWidgetGetNoShowAll}
\description{Returns the current value of the GtkWidget:no-show-all property, 
which determines whether calls to \code{\link{gtkWidgetShowAll}} and 
\code{\link{gtkWidgetHideAll}} will affect this widget.}
\usage{gtkWidgetGetNoShowAll(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.4}
\value{[logical] the current value of the "no-show-all" property.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
