\alias{gtkWidgetGetParent}
\name{gtkWidgetGetParent}
\title{gtkWidgetGetParent}
\description{Returns the parent container of \code{widget}.}
\usage{gtkWidgetGetParent(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkWidget}}]  a \code{\link{GtkWidget}}}}
\value{[\code{\link{GtkWidget}}]  the parent container of \code{widget}, or \code{NULL}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
