\alias{gtkWidgetGetParentWindow}
\name{gtkWidgetGetParentWindow}
\title{gtkWidgetGetParentWindow}
\description{Gets \code{widget}'s parent window.}
\usage{gtkWidgetGetParentWindow(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}.}}
\value{[\verb{\link{GdkWindow}}]  the parent window of \code{widget}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
