\alias{gtkWidgetGetRealized}
\name{gtkWidgetGetRealized}
\title{gtkWidgetGetRealized}
\description{Determines whether \code{widget} is realized.}
\usage{gtkWidgetGetRealized(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.20}
\value{[logical] \code{TRUE} if \code{widget} is realized, \code{FALSE} otherwise}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
