\alias{gtkWidgetGetState}
\name{gtkWidgetGetState}
\title{gtkWidgetGetState}
\description{Returns the widget's state. See \code{\link{gtkWidgetSetState}}.}
\usage{gtkWidgetGetState(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.18}
\value{[\code{\link{GtkStateType}}] the state of \code{widget}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
