\alias{gtkWidgetGetTooltipMarkup}
\name{gtkWidgetGetTooltipMarkup}
\title{gtkWidgetGetTooltipMarkup}
\description{Gets the contents of the tooltip for \code{widget}.}
\usage{gtkWidgetGetTooltipMarkup(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.12}
\value{[character] the tooltip text, or \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
