\alias{gtkWidgetGetTooltipWindow}
\name{gtkWidgetGetTooltipWindow}
\title{gtkWidgetGetTooltipWindow}
\description{Returns the \verb{\link{GtkWindow}} of the current tooltip. This can be the
GtkWindow created by default, or the custom tooltip window set
using \code{\link{gtkWidgetSetTooltipWindow}}.}
\usage{gtkWidgetGetTooltipWindow(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}}}
\details{  Since  2.12}
\value{[\verb{\link{GtkWindow}}]  The \verb{\link{GtkWindow}} of the current tooltip.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
