\alias{gtkWidgetGetVisible}
\name{gtkWidgetGetVisible}
\title{gtkWidgetGetVisible}
\description{Determines whether the widget is visible. Note that this doesn't
take into account whether the widget's parent is also visible
or the widget is obscured in any way.}
\usage{gtkWidgetGetVisible(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{See \code{\link{gtkWidgetSetVisible}}.
  Since 2.18}
\value{[logical] \code{TRUE} if the widget is visible}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
