\alias{gtkWidgetGetVisual}
\name{gtkWidgetGetVisual}
\title{gtkWidgetGetVisual}
\description{Gets the visual that will be used to render \code{widget}.}
\usage{gtkWidgetGetVisual(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}}}
\value{[\verb{\link{GdkVisual}}]  the visual for \code{widget}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
