\alias{gtkWidgetGetWindow}
\name{gtkWidgetGetWindow}
\title{gtkWidgetGetWindow}
\description{Returns the widget's window if it is realized, \code{NULL} otherwise}
\usage{gtkWidgetGetWindow(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.14}
\value{[\code{\link{GdkWindow}}] \code{widget}'s window.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
