\alias{gtkWidgetHasDefault}
\name{gtkWidgetHasDefault}
\title{gtkWidgetHasDefault}
\description{Determines whether \code{widget} is the current default widget within its
toplevel. See \code{\link{gtkWidgetSetCanDefault}}.}
\usage{gtkWidgetHasDefault(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.18}
\value{[logical] \code{TRUE} if \code{widget} is the current default widget within
its toplevel, \code{FALSE} otherwise}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
