\alias{gtkWidgetHasFocus}
\name{gtkWidgetHasFocus}
\title{gtkWidgetHasFocus}
\description{Determines if the widget has the global input focus. See
\code{\link{gtkWidgetIsFocus}} for the difference between having the global
input focus, and only having the focus within a toplevel.}
\usage{gtkWidgetHasFocus(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.18}
\value{[logical] \code{TRUE} if the widget has the global input focus.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
