\alias{gtkWidgetHasRcStyle}
\name{gtkWidgetHasRcStyle}
\title{gtkWidgetHasRcStyle}
\description{Determines if the widget style has been looked up through the rc mechanism.}
\usage{gtkWidgetHasRcStyle(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.20}
\value{[logical] \code{TRUE} if the widget has been looked up through the rc
mechanism, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
