\alias{gtkWidgetHide}
\name{gtkWidgetHide}
\title{gtkWidgetHide}
\description{Reverses the effects of \code{\link{gtkWidgetShow}}, causing the widget to be
hidden (invisible to the user).}
\usage{gtkWidgetHide(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkWidget}}]  a \code{\link{GtkWidget}}}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
