\alias{gtkWidgetIsDrawable}
\name{gtkWidgetIsDrawable}
\title{gtkWidgetIsDrawable}
\description{Determines whether \code{widget} can be drawn to. A widget can be drawn
to if it is mapped and visible.}
\usage{gtkWidgetIsDrawable(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.18}
\value{[logical] \code{TRUE} if \code{widget} is drawable, \code{FALSE} otherwise}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
