\alias{gtkWidgetIsSensitive}
\name{gtkWidgetIsSensitive}
\title{gtkWidgetIsSensitive}
\description{Reports whether the widget is sensitive.}
\usage{
gtkWidgetIsSensitive(wid)
}
\arguments{
\item{wid}{The widget to check for sensitivity.}
}
\value{
boolean indicating whether the widget is sensitive
}
\author{Michael Lawrence}
\keyword{internal}
\keyword{interface}
