\alias{gtkWidgetMap}
\name{gtkWidgetMap}
\title{gtkWidgetMap}
\description{This function is only for use in widget implementations. Causes
a widget to be mapped if it isn't already.}
\usage{gtkWidgetMap(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkWidget}}]  a \code{\link{GtkWidget}}}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
