\alias{gtkWidgetPopCompositeChild}
\name{gtkWidgetPopCompositeChild}
\title{gtkWidgetPopCompositeChild}
\description{Cancels the effect of a previous call to \code{\link{gtkWidgetPushCompositeChild}}.}
\usage{gtkWidgetPopCompositeChild()}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
