\alias{gtkWidgetQueueClear}
\name{gtkWidgetQueueClear}
\title{gtkWidgetQueueClear}
\description{
This function does the same as \code{\link{gtkWidgetQueueDraw}}.
\strong{WARNING: \code{gtk_widget_queue_clear} is deprecated and should not be used in newly-written code. Use \code{\link{gtkWidgetQueueDraw}} instead.}
}
\usage{gtkWidgetQueueClear(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
