\alias{gtkWidgetQueueDraw}
\name{gtkWidgetQueueDraw}
\title{gtkWidgetQueueDraw}
\description{Equivalent to calling \code{\link{gtkWidgetQueueDrawArea}} for the
entire area of a widget.}
\usage{gtkWidgetQueueDraw(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkWidget}}]  a \code{\link{GtkWidget}}}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
