\alias{gtkWidgetQueueResizeNoRedraw}
\name{gtkWidgetQueueResizeNoRedraw}
\title{gtkWidgetQueueResizeNoRedraw}
\description{This function works like \code{\link{gtkWidgetQueueResize}}, 
except that the widget is not invalidated.}
\usage{gtkWidgetQueueResizeNoRedraw(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
