\alias{gtkWidgetResetRcStyles}
\name{gtkWidgetResetRcStyles}
\title{gtkWidgetResetRcStyles}
\description{Reset the styles of \code{widget} and all descendents, so when
they are looked up again, they get the correct values
for the currently loaded RC file settings.}
\usage{gtkWidgetResetRcStyles(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}.}}
\details{This function is not useful for applications.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
