\alias{gtkWidgetResetShapes}
\name{gtkWidgetResetShapes}
\title{gtkWidgetResetShapes}
\description{Recursively resets the shape on this widget and its descendants.}
\usage{gtkWidgetResetShapes(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkWidget}}]  a \code{\link{GtkWidget}}}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
