\alias{gtkWidgetSet}
\name{gtkWidgetSet}
\title{gtkWidgetSet}
\description{
Like \code{\link{gObjectSet}} - there's no reason to use this instead of
\code{\link{gObjectSet}}.
\strong{WARNING: \code{gtk_widget_set} is deprecated and should not be used in newly-written code.}
}
\usage{gtkWidgetSet(obj, ...)}
\arguments{\item{\code{...}}{[character]  name of first property to set}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
