\alias{gtkWidgetSetAllocation}
\name{gtkWidgetSetAllocation}
\title{gtkWidgetSetAllocation}
\description{Sets the widget's allocation.  This should not be used
directly, but from within a widget's size_allocate method.}
\usage{gtkWidgetSetAllocation(object, allocation)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWidget}}}
\item{\verb{allocation}}{a pointer to a \code{\link{GtkAllocation}} to copy from}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
