\alias{gtkWidgetSetCanDefault}
\name{gtkWidgetSetCanDefault}
\title{gtkWidgetSetCanDefault}
\description{Specifies whether \code{widget} can be a default widget. See
\code{\link{gtkWidgetGrabDefault}} for details about the meaning of
"default".}
\usage{gtkWidgetSetCanDefault(object, can.default)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWidget}}}
\item{\verb{can.default}}{whether or not \code{widget} can be a default widget.}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
