\alias{gtkWidgetSetCanFocus}
\name{gtkWidgetSetCanFocus}
\title{gtkWidgetSetCanFocus}
\description{Specifies whether \code{widget} can own the input focus. See
\code{\link{gtkWidgetGrabFocus}} for actually setting the input focus on a
widget.}
\usage{gtkWidgetSetCanFocus(object, can.focus)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWidget}}}
\item{\verb{can.focus}}{whether or not \code{widget} can own the input focus.}
}
\details{Since 2.18}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
