\alias{gtkWidgetSetDefaultDirection}
\name{gtkWidgetSetDefaultDirection}
\title{gtkWidgetSetDefaultDirection}
\description{Sets the default reading direction for widgets where the
direction has not been explicitly set by \code{\link{gtkWidgetSetDirection}}.}
\usage{gtkWidgetSetDefaultDirection(dir)}
\arguments{\item{\code{dir}}{[\code{\link{GtkTextDirection}}]  the new default direction. This cannot be
       \code{GTK_TEXT_DIR_NONE}.}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
