\alias{gtkWidgetSetExtensionEvents}
\name{gtkWidgetSetExtensionEvents}
\title{gtkWidgetSetExtensionEvents}
\description{Sets the extension events mask to \code{mode}. See \code{\link{GdkExtensionMode}}
and \code{\link{gdkInputSetExtensionEvents}}.}
\usage{gtkWidgetSetExtensionEvents(object, mode)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}}
\item{\verb{mode}}{[\code{\link{GdkExtensionMode}}]  bitfield of extension events to receive}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
