\alias{gtkWidgetSetHasTooltip}
\name{gtkWidgetSetHasTooltip}
\title{gtkWidgetSetHasTooltip}
\description{Sets the has-tooltip property on \code{widget} to \code{has.tooltip}.  See
GtkWidget:has-tooltip for more information.}
\usage{gtkWidgetSetHasTooltip(object, has.tooltip)}
\arguments{
\item{\code{object}}{[\code{\link{GtkWidget}}]  a \code{\link{GtkWidget}}}
\item{\code{has.tooltip}}{[logical]  whether or not \code{widget} has a tooltip.}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
