\alias{gtkWidgetSetParentWindow}
\name{gtkWidgetSetParentWindow}
\title{gtkWidgetSetParentWindow}
\description{Sets a non default parent window for \code{widget}.}
\usage{gtkWidgetSetParentWindow(object, parent.window)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}.}
\item{\verb{parent.window}}{[\verb{\link{GdkWindow}}]  the new parent window.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
