\alias{gtkWidgetSetSensitive}
\name{gtkWidgetSetSensitive}
\title{gtkWidgetSetSensitive}
\description{Sets the sensitivity of a widget. A widget is sensitive if the user
can interact with it. Insensitive widgets are "grayed out" and the
user can't interact with them. Insensitive widgets are known as
"inactive", "disabled", or "ghosted" in some other toolkits.}
\usage{gtkWidgetSetSensitive(object, sensitive)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWidget}}}
\item{\verb{sensitive}}{\code{TRUE} to make the widget sensitive}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
