\alias{gtkWidgetSizeAllocate}
\name{gtkWidgetSizeAllocate}
\title{gtkWidgetSizeAllocate}
\description{This function is only used by \verb{\link{GtkContainer}} subclasses, to assign a size
and position to their child widgets.}
\usage{gtkWidgetSizeAllocate(object, allocation)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}}
\item{\verb{allocation}}{[\verb{\link{GtkAllocation}}]  position and size to be allocated to \code{widget}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
