\alias{gtkWidgetSavedState}
\alias{gtkWidgetState}
\name{gtkWidgetState}
\title{gtkWidgetState}
\description{Get the states of a widget.}
\usage{
gtkWidgetState(wid)
gtkWidgetSavedState(wid)
}
\arguments{
\item{wid}{The widget to query for its (saved) state.}
}
\value{
A state (\code{\link{GtkStateType}}) of the widget, either saved or current.
}
\author{Michael Lawrence}
\keyword{internal}
\keyword{interface}
