\alias{gtkWidgetStyleGet}
\name{gtkWidgetStyleGet}
\title{gtkWidgetStyleGet}
\description{Gets the values of a multiple style properties of \code{widget}.}
\usage{gtkWidgetStyleGet(object, ...)}
\arguments{
\item{\code{object}}{[\code{\link{GtkWidget}}]  a \code{\link{GtkWidget}}}
\item{\code{...}}{ pairs of property names and locations to 
  return the property values, starting with the location for 
  \code{first.property.name}, terminated by \code{NULL}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
