\alias{gtkWidgetTriggerTooltipQuery}
\name{gtkWidgetTriggerTooltipQuery}
\title{gtkWidgetTriggerTooltipQuery}
\description{Triggers a tooltip query on the display where the toplevel of \code{widget}
is located. See \code{\link{gtkTooltipTriggerTooltipQuery}} for more
information.}
\usage{gtkWidgetTriggerTooltipQuery(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
