\alias{gtkWidgetUnparent}
\name{gtkWidgetUnparent}
\title{gtkWidgetUnparent}
\description{This function is only for use in widget implementations.
Should be called by implementations of the remove method
on \code{\link{GtkContainer}}, to dissociate a child from the container.}
\usage{gtkWidgetUnparent(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
