\alias{gtkWidgetUnrealize}
\name{gtkWidgetUnrealize}
\title{gtkWidgetUnrealize}
\description{This function is only useful in widget implementations.
Causes a widget to be unrealized (frees all GDK resources
associated with the widget, such as \code{widget->window}).}
\usage{gtkWidgetUnrealize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
