\alias{gtkWindowActivateFocus}
\name{gtkWindowActivateFocus}
\title{gtkWindowActivateFocus}
\description{Activates the current focused widget within the window.}
\usage{gtkWindowActivateFocus(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\value{[logical] \code{TRUE} if a widget got activated.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
