\alias{gtkWindowAddMnemonic}
\name{gtkWindowAddMnemonic}
\title{gtkWindowAddMnemonic}
\description{Adds a mnemonic to this window.}
\usage{gtkWindowAddMnemonic(object, keyval, target)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkWindow}}]  a \verb{\link{GtkWindow}}}
\item{\verb{keyval}}{[numeric]  the mnemonic}
\item{\verb{target}}{[\verb{\link{GtkWidget}}]  the widget that gets activated by the mnemonic}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
