\alias{gtkWindowGetAcceptFocus}
\name{gtkWindowGetAcceptFocus}
\title{gtkWindowGetAcceptFocus}
\description{Gets the value set by \code{\link{gtkWindowSetAcceptFocus}}.}
\usage{gtkWindowGetAcceptFocus(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if window should receive the input focus}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
