\alias{gtkWindowGetDecorated}
\name{gtkWindowGetDecorated}
\title{gtkWindowGetDecorated}
\description{Returns whether the window has been set to have decorations
such as a title bar via \code{\link{gtkWindowSetDecorated}}.}
\usage{gtkWindowGetDecorated(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\value{[logical] \code{TRUE} if the window has been set to have decorations}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
