\alias{gtkWindowGetDefaultIconList}
\name{gtkWindowGetDefaultIconList}
\title{gtkWindowGetDefaultIconList}
\description{Gets the value set by \code{\link{gtkWindowSetDefaultIconList}}.}
\usage{gtkWindowGetDefaultIconList()}
\value{[list]  copy of default icon list }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
