\alias{gtkWindowGetDefaultIconName}
\name{gtkWindowGetDefaultIconName}
\title{gtkWindowGetDefaultIconName}
\description{Returns the fallback icon name for windows that has been set
with \code{\link{gtkWindowSetDefaultIconName}}. It
is only valid until the next call to
\code{\link{gtkWindowSetDefaultIconName}}.}
\usage{gtkWindowGetDefaultIconName()}
\details{Since 2.16}
\value{[character] the fallback icon name for windows}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
