\alias{gtkWindowGetDefaultWidget}
\name{gtkWindowGetDefaultWidget}
\title{gtkWindowGetDefaultWidget}
\description{Returns the default widget for \code{window}. See \code{\link{gtkWindowSetDefault}}
for more details.}
\usage{gtkWindowGetDefaultWidget(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\details{Since 2.14}
\value{[\code{\link{GtkWidget}}] the default widget, or \code{NULL} if there is none.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
