\alias{gtkWindowGetDeletable}
\name{gtkWindowGetDeletable}
\title{gtkWindowGetDeletable}
\description{Returns whether the window has been set to have a close button
via \code{\link{gtkWindowSetDeletable}}.}
\usage{gtkWindowGetDeletable(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\details{Since 2.10}
\value{[logical] \code{TRUE} if the window has been set to have a close button}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
