\alias{gtkWindowGetDestroyWithParent}
\name{gtkWindowGetDestroyWithParent}
\title{gtkWindowGetDestroyWithParent}
\description{Returns whether the window will be destroyed with its transient parent. See
\code{\link{gtkWindowSetDestroyWithParent}}.}
\usage{gtkWindowGetDestroyWithParent(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\value{[logical] \code{TRUE} if the window will be destroyed with its transient parent.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
