\alias{gtkWindowGetFocusOnMap}
\name{gtkWindowGetFocusOnMap}
\title{gtkWindowGetFocusOnMap}
\description{Gets the value set by \code{\link{gtkWindowSetFocusOnMap}}.}
\usage{gtkWindowGetFocusOnMap(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\details{Since 2.6}
\value{[logical] \code{TRUE} if window should receive the input focus when
mapped.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
