\alias{gtkWindowGetHasFrame}
\name{gtkWindowGetHasFrame}
\title{gtkWindowGetHasFrame}
\description{Accessor for whether the window has a frame window exterior to
\code{window->window}. Gets the value set by \code{\link{gtkWindowSetHasFrame}}.}
\usage{gtkWindowGetHasFrame(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\value{[logical] \code{TRUE} if a frame has been added to the window
via \code{\link{gtkWindowSetHasFrame}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
