\alias{gtkWindowGetIcon}
\name{gtkWindowGetIcon}
\title{gtkWindowGetIcon}
\description{Gets the value set by \code{\link{gtkWindowSetIcon}} (or if you've
called \code{\link{gtkWindowSetIconList}}, gets the first icon in
the icon list).}
\usage{gtkWindowGetIcon(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkWindow}}]  a \verb{\link{GtkWindow}}}}
\value{[\verb{\link{GdkPixbuf}}]  icon for window}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
