\alias{gtkWindowGetIconList}
\name{gtkWindowGetIconList}
\title{gtkWindowGetIconList}
\description{Retrieves the list of icons set by \code{\link{gtkWindowSetIconList}}.
The list is copied, but the reference count on each
member won't be incremented.}
\usage{gtkWindowGetIconList(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkWindow}}]  a \verb{\link{GtkWindow}}}}
\value{[list]  copy of window's icon list}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
