\alias{gtkWindowGetIconName}
\name{gtkWindowGetIconName}
\title{gtkWindowGetIconName}
\description{Returns the name of the themed icon for the window,
see \code{\link{gtkWindowSetIconName}}.}
\usage{gtkWindowGetIconName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\details{Since 2.6}
\value{[character] the icon name or \code{NULL} if the window has
no themed icon}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
