\alias{gtkWindowGetMnemonicModifier}
\name{gtkWindowGetMnemonicModifier}
\title{gtkWindowGetMnemonicModifier}
\description{Returns the mnemonic modifier for this window. See
\code{\link{gtkWindowSetMnemonicModifier}}.}
\usage{gtkWindowGetMnemonicModifier(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\value{[\code{\link{GdkModifierType}}] the modifier mask used to activate
mnemonics on this window.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
