\alias{gtkWindowGetMnemonicsVisible}
\name{gtkWindowGetMnemonicsVisible}
\title{gtkWindowGetMnemonicsVisible}
\description{\emph{undocumented
}}
\usage{gtkWindowGetMnemonicsVisible(object)}
\arguments{\item{\verb{object}}{\emph{undocumented
}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
