\alias{gtkWindowGetOpacity}
\name{gtkWindowGetOpacity}
\title{gtkWindowGetOpacity}
\description{Fetches the requested opacity for this window. See
\code{\link{gtkWindowSetOpacity}}.}
\usage{gtkWindowGetOpacity(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\details{Since 2.12}
\value{[numeric] the requested opacity for this window.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
