\alias{gtkWindowGetRole}
\name{gtkWindowGetRole}
\title{gtkWindowGetRole}
\description{Returns the role of the window. See \code{\link{gtkWindowSetRole}} for
further explanation.}
\usage{gtkWindowGetRole(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\value{[character] the role of the window if set, or \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
