\alias{gtkWindowGetScreen}
\name{gtkWindowGetScreen}
\title{gtkWindowGetScreen}
\description{Returns the \verb{\link{GdkScreen}} associated with \code{window}.}
\usage{gtkWindowGetScreen(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkWindow}}]  a \verb{\link{GtkWindow}}.}}
\details{  Since  2.2}
\value{[\verb{\link{GdkScreen}}]  a \verb{\link{GdkScreen}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
