\alias{gtkWindowGetSkipPagerHint}
\name{gtkWindowGetSkipPagerHint}
\title{gtkWindowGetSkipPagerHint}
\description{Gets the value set by \code{\link{gtkWindowSetSkipPagerHint}}.}
\usage{gtkWindowGetSkipPagerHint(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkWindow}}]  a \code{\link{GtkWindow}}}}
\details{  Since  2.2}
\value{[logical] \code{TRUE} if window shouldn't be in pager}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
