\alias{gtkWindowGetSkipTaskbarHint}
\name{gtkWindowGetSkipTaskbarHint}
\title{gtkWindowGetSkipTaskbarHint}
\description{Gets the value set by \code{\link{gtkWindowSetSkipTaskbarHint}}}
\usage{gtkWindowGetSkipTaskbarHint(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkWindow}}]  a \code{\link{GtkWindow}}}}
\details{  Since  2.2}
\value{[logical]  \code{TRUE} if window shouldn't be in taskbar}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
