\alias{gtkWindowGetTitle}
\name{gtkWindowGetTitle}
\title{gtkWindowGetTitle}
\description{Retrieves the title of the window. See \code{\link{gtkWindowSetTitle}}.}
\usage{gtkWindowGetTitle(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\value{[character] the title of the window, or \code{NULL} if none has
been set explicitely.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
