\alias{gtkWindowGetWindowType}
\name{gtkWindowGetWindowType}
\title{gtkWindowGetWindowType}
\description{Gets the type of the window. See \code{\link{GtkWindowType}}.}
\usage{gtkWindowGetWindowType(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\details{Since 2.20}
\value{[\code{\link{GtkWindowType}}] the type of the window}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
