\alias{GtkWindowGroup}
\alias{gtkWindowGroup}
\name{GtkWindowGroup}
\title{GtkWindowGroup}
\description{Limit the effect of grabs}
\usage{
\link{gtkWindowGroupNew}()
\link{gtkWindowGroupAddWindow}(object, window)
\link{gtkWindowGroupRemoveWindow}(object, window)
gtkWindowGroup()
}
\section{Hierarchy}{\preformatted{  \link{GObject}
   +----GtkWindowGroup}}
\section{Structures}{\describe{\item{\code{GtkWindowGroup}}{\emph{undocumented
}}}}
\section{Convenient Construction}{\code{gtkWindowGroup} is the equivalent of \code{\link{gtkWindowGroupNew}}.}
\references{\url{http://developer.gnome.org/doc/API/2.0/gtk/GtkWindowGroup.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
