\alias{gtkWindowGroupListWindows}
\name{gtkWindowGroupListWindows}
\title{gtkWindowGroupListWindows}
\description{Returns a list of the \verb{GtkWindows} that belong to \code{window.group}.}
\usage{gtkWindowGroupListWindows(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindowGroup}}}}
\details{Since 2.14}
\value{[list] A newly-allocated list of
windows inside the group. \emph{[  \acronym{element-type}   GtkWidget][  \acronym{transfer container}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
