\alias{gtkWindowGroupNew}
\name{gtkWindowGroupNew}
\title{gtkWindowGroupNew}
\description{Creates a new \verb{\link{GtkWindowGroup}} object. Grabs added with
\code{\link{gtkGrabAdd}} only affect windows within the same \verb{\link{GtkWindowGroup}}.}
\usage{gtkWindowGroupNew()}
\value{[\verb{\link{GtkWindowGroup}}]  a new \verb{\link{GtkWindowGroup}}. }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
