\alias{gtkWindowGroupRemoveWindow}
\name{gtkWindowGroupRemoveWindow}
\title{gtkWindowGroupRemoveWindow}
\description{Removes a window from a \code{\link{GtkWindowGroup}}.}
\usage{gtkWindowGroupRemoveWindow(object, window)}
\arguments{
\item{\code{object}}{[\code{\link{GtkWindowGroup}}]  a \code{\link{GtkWindowGroup}}}
\item{\code{window}}{[\code{\link{GtkWindow}}]  the \code{\link{GtkWindow}} to remove}
}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
