\alias{gtkWindowMnemonicActivate}
\name{gtkWindowMnemonicActivate}
\title{gtkWindowMnemonicActivate}
\description{Activates the targets associated with the mnemonic.}
\usage{gtkWindowMnemonicActivate(object, keyval, modifier)}
\arguments{
\item{\code{object}}{[\code{\link{GtkWindow}}]  a \code{\link{GtkWindow}}}
\item{\code{keyval}}{[numeric]  the mnemonic}
\item{\code{modifier}}{[\code{\link{GdkModifierType}}]  the modifiers }
}
\value{[logical]  \code{TRUE} if the activation is done. }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
