\alias{gtkWindowRemoveMnemonic}
\name{gtkWindowRemoveMnemonic}
\title{gtkWindowRemoveMnemonic}
\description{Removes a mnemonic from this window.}
\usage{gtkWindowRemoveMnemonic(object, keyval, target)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkWindow}}]  a \verb{\link{GtkWindow}}}
\item{\verb{keyval}}{[numeric]  the mnemonic}
\item{\verb{target}}{[\verb{\link{GtkWidget}}]  the widget that gets activated by the mnemonic}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
