\alias{gtkWindowReshowWithInitialSize}
\name{gtkWindowReshowWithInitialSize}
\title{gtkWindowReshowWithInitialSize}
\description{Hides \code{window}, then reshows it, resetting the
default size and position of the window. Used
by GUI builders only.}
\usage{gtkWindowReshowWithInitialSize(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkWindow}}]  a \code{\link{GtkWindow}}}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
