\alias{gtkWindowSetDefaultIcon}
\name{gtkWindowSetDefaultIcon}
\title{gtkWindowSetDefaultIcon}
\description{Sets an icon to be used as fallback for windows that haven't
had \code{\link{gtkWindowSetIcon}} called on them from a pixbuf.}
\usage{gtkWindowSetDefaultIcon(icon)}
\arguments{\item{\verb{icon}}{the icon}}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
