\alias{gtkWindowSetDefaultIconName}
\name{gtkWindowSetDefaultIconName}
\title{gtkWindowSetDefaultIconName}
\description{Sets an icon to be used as fallback for windows that haven't
had \code{\link{gtkWindowSetIconList}} called on them from a named
themed icon, see \code{\link{gtkWindowSetIconName}}.}
\usage{gtkWindowSetDefaultIconName(name)}
\arguments{\item{\verb{name}}{the name of the themed icon}}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
