\alias{gtkWindowSetFocusOnMap}
\name{gtkWindowSetFocusOnMap}
\title{gtkWindowSetFocusOnMap}
\description{Windows may set a hint asking the desktop environment not to receive
the input focus when the window is mapped.  This function sets this
hint.}
\usage{gtkWindowSetFocusOnMap(object, setting)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWindow}}}
\item{\verb{setting}}{\code{TRUE} to let this window receive input focus on map}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
