\alias{gtkWindowSetMnemonicsVisible}
\name{gtkWindowSetMnemonicsVisible}
\title{gtkWindowSetMnemonicsVisible}
\description{Sets the \verb{"mnemonics-visible"} property.}
\usage{gtkWindowSetMnemonicsVisible(object, setting)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWindow}}}
\item{\verb{setting}}{the new value}
}
\details{Since 2.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
