\alias{gtkWindowSetPosition}
\name{gtkWindowSetPosition}
\title{gtkWindowSetPosition}
\description{Sets a position constraint for this window. If the old or new
constraint is \code{GTK_WIN_POS_CENTER_ALWAYS}, this will also cause
the window to be repositioned to satisfy the new constraint.}
\usage{gtkWindowSetPosition(object, position)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWindow}}.}
\item{\verb{position}}{a position constraint.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
