\alias{gtkWindowSetResizable}
\name{gtkWindowSetResizable}
\title{gtkWindowSetResizable}
\description{Sets whether the user can resize a window. Windows are user resizable
by default.}
\usage{gtkWindowSetResizable(object, resizable)}
\arguments{
\item{\code{object}}{[\code{\link{GtkWindow}}]  a \code{\link{GtkWindow}}}
\item{\code{resizable}}{[logical]  \code{TRUE} if the user can resize this window}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
