\alias{gtkWindowSetScreen}
\name{gtkWindowSetScreen}
\title{gtkWindowSetScreen}
\description{Sets the \verb{\link{GdkScreen}} where the \code{window} is displayed; if
the window is already mapped, it will be unmapped, and
then remapped on the new screen.}
\usage{gtkWindowSetScreen(object, screen)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkWindow}}]  a \verb{\link{GtkWindow}}.}
\item{\verb{screen}}{[\verb{\link{GdkScreen}}]  a \verb{\link{GdkScreen}}.}
}
\details{  Since  2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
