\alias{gtkWindowSetSkipTaskbarHint}
\name{gtkWindowSetSkipTaskbarHint}
\title{gtkWindowSetSkipTaskbarHint}
\description{Windows may set a hint asking the desktop environment not to display
the window in the task bar. This function sets this hint.}
\usage{gtkWindowSetSkipTaskbarHint(object, setting)}
\arguments{
\item{\code{object}}{[\code{\link{GtkWindow}}]  a \code{\link{GtkWindow}} }
\item{\code{setting}}{[logical]  \code{TRUE} to keep this window from appearing in the task bar}
}
\details{  Since  2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
