\alias{gtkWindowSetTitle}
\name{gtkWindowSetTitle}
\title{gtkWindowSetTitle}
\description{Sets the title of the \code{\link{GtkWindow}}. The title of a window will be
displayed in its title bar; on the X Window System, the title bar
is rendered by the window
manager, so exactly how the title appears to users may vary
according to a user's exact configuration. The title should help a
user distinguish this window from other windows they may have
open. A good title might include the application name and current
document filename, for example.}
\usage{gtkWindowSetTitle(object, title)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWindow}}}
\item{\verb{title}}{title of the window}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
