\alias{GVfs}
\name{GVfs}
\title{GVfs}
\description{Virtual File System}
\section{Methods and Functions}{
\code{\link{gVfsGetFileForPath}(object, path)}\cr
\code{\link{gVfsGetFileForUri}(object, uri)}\cr
\code{\link{gVfsParseName}(object, parse.name)}\cr
\code{\link{gVfsGetDefault}()}\cr
\code{\link{gVfsGetLocal}()}\cr
\code{\link{gVfsIsActive}(object)}\cr
\code{\link{gVfsGetSupportedUriSchemes}(object)}\cr

}
\section{Hierarchy}{\preformatted{GObject
   +----GVfs}}
\section{Detailed Description}{Entry point for using GIO functionality.}
\section{Structures}{\describe{\item{\verb{GVfs}}{
Virtual File System object.

}}}
\references{\url{https://developer-old.gnome.org/gio/stable/GVfs.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
