\alias{gVfsGetFileForPath}
\name{gVfsGetFileForPath}
\title{gVfsGetFileForPath}
\description{Gets a \code{\link{GFile}} for \code{path}.}
\usage{gVfsGetFileForPath(object, path)}
\arguments{
\item{\verb{object}}{a \code{\link{GVfs}}.}
\item{\verb{path}}{a string containing a VFS path.}
}
\value{[\code{\link{GFile}}] a \code{\link{GFile}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
