\alias{gVfsGetSupportedUriSchemes}
\name{gVfsGetSupportedUriSchemes}
\title{gVfsGetSupportedUriSchemes}
\description{Gets a list of URI schemes supported by \code{vfs}.}
\usage{gVfsGetSupportedUriSchemes(object)}
\arguments{\item{\verb{object}}{a \code{\link{GVfs}}.}}
\value{[character] a list of strings.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
