\alias{gVfsIsActive}
\name{gVfsIsActive}
\title{gVfsIsActive}
\description{Checks if the VFS is active.}
\usage{gVfsIsActive(object)}
\arguments{\item{\verb{object}}{a \code{\link{GVfs}}.}}
\value{[logical] \code{TRUE} if construction of the \code{vfs} was successful and it is now active.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
