\alias{gVolumeCanEject}
\name{gVolumeCanEject}
\title{gVolumeCanEject}
\description{Checks if a volume can be ejected.}
\usage{gVolumeCanEject(object)}
\arguments{\item{\verb{object}}{a \code{\link{GVolume}}.}}
\value{[logical] \code{TRUE} if the \code{volume} can be ejected. \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
