\alias{gVolumeCanMount}
\name{gVolumeCanMount}
\title{gVolumeCanMount}
\description{Checks if a volume can be mounted.}
\usage{gVolumeCanMount(object)}
\arguments{\item{\verb{object}}{a \code{\link{GVolume}}.}}
\value{[logical] \code{TRUE} if the \code{volume} can be mounted. \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
