\alias{gVolumeEnumerateIdentifiers}
\name{gVolumeEnumerateIdentifiers}
\title{gVolumeEnumerateIdentifiers}
\description{Gets the kinds of identifiers
that \code{volume} has. Use \code{gVolumeGetIdentifer()} to obtain 
the identifiers themselves.}
\usage{gVolumeEnumerateIdentifiers(object)}
\arguments{\item{\verb{object}}{a \code{\link{GVolume}}}}
\value{[char] a list of strings containing
kinds of identifiers.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
