\alias{gVolumeGetDrive}
\name{gVolumeGetDrive}
\title{gVolumeGetDrive}
\description{Gets the drive for the \code{volume}.}
\usage{gVolumeGetDrive(object)}
\arguments{\item{\verb{object}}{a \code{\link{GVolume}}.}}
\value{[\code{\link{GDrive}}] a \code{\link{GDrive}} or \code{NULL} if \code{volume} is not associated with a drive.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
