\alias{gVolumeGetIcon}
\name{gVolumeGetIcon}
\title{gVolumeGetIcon}
\description{Gets the icon for \code{volume}.}
\usage{gVolumeGetIcon(object)}
\arguments{\item{\verb{object}}{a \code{\link{GVolume}}.}}
\value{[\code{\link{GIcon}}] a \code{\link{GIcon}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
