\alias{gVolumeGetMount}
\name{gVolumeGetMount}
\title{gVolumeGetMount}
\description{Gets the mount for the \code{volume}.}
\usage{gVolumeGetMount(object)}
\arguments{\item{\verb{object}}{a \code{\link{GVolume}}.}}
\value{[\code{\link{GMount}}] a \code{\link{GMount}} or \code{NULL} if \code{volume} isn't mounted.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
