\alias{gVolumeGetName}
\name{gVolumeGetName}
\title{gVolumeGetName}
\description{Gets the name of \code{volume}.}
\usage{gVolumeGetName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GVolume}}.}}
\value{[char] the name for the given \code{volume}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
