\alias{gVolumeMonitorGet}
\name{gVolumeMonitorGet}
\title{gVolumeMonitorGet}
\description{Gets the volume monitor used by gio.}
\usage{gVolumeMonitorGet()}
\value{[\code{\link{GVolumeMonitor}}] a reference to the \code{\link{GVolumeMonitor}} used by gio.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
