\alias{gVolumeMonitorGetConnectedDrives}
\name{gVolumeMonitorGetConnectedDrives}
\title{gVolumeMonitorGetConnectedDrives}
\description{Gets a list of drives connected to the system.}
\usage{gVolumeMonitorGetConnectedDrives(object)}
\arguments{\item{\verb{object}}{a \code{\link{GVolumeMonitor}}.}}
\value{[list] a \verb{list} of connected \code{\link{GDrive}} objects.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
