\alias{gVolumeMonitorGetMounts}
\name{gVolumeMonitorGetMounts}
\title{gVolumeMonitorGetMounts}
\description{Gets a list of the mounts on the system.}
\usage{gVolumeMonitorGetMounts(object)}
\arguments{\item{\verb{object}}{a \code{\link{GVolumeMonitor}}.}}
\value{[list] a \verb{list} of \code{\link{GMount}} objects.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
