\alias{gVolumeMonitorGetVolumes}
\name{gVolumeMonitorGetVolumes}
\title{gVolumeMonitorGetVolumes}
\description{Gets a list of the volumes on the system.}
\usage{gVolumeMonitorGetVolumes(object)}
\arguments{\item{\verb{object}}{a \code{\link{GVolumeMonitor}}.}}
\value{[list] a \verb{list} of \code{\link{GVolume}} objects.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
