\alias{gVolumeShouldAutomount}
\name{gVolumeShouldAutomount}
\title{gVolumeShouldAutomount}
\description{Returns whether the volume should be automatically mounted.}
\usage{gVolumeShouldAutomount(object)}
\arguments{\item{\verb{object}}{a \code{\link{GVolume}}}}
\value{[logical] \code{TRUE} if the volume should be automatically mounted.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
