\alias{Libglade}
\name{Libglade}
\title{Libglade}
\description{Libglade loads and parses XML descriptions of user interfaces at runtime. It also provides functions that can be used to connect signal handlers to parts of the interface.}
\details{
The RGtk binding to the Libglade library consists of the following components:
\describe{\item{\link{GladeXML}}{Allows dynamic loading of user interfaces from XML descriptions.}}
}
\references{\url{http://library.gnome.org/devel//libglade}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{interface}
