\alias{pango-ATSUI-Fonts}
\name{pango-ATSUI-Fonts}
\title{ATSUI Fonts}
\description{Font handling and rendering with ATSUI fonts}
\section{Detailed Description}{The functions and functions in this section are used to access fonts natively on
MacOS systems using the ATSUI text rendering subsystem.}
\references{\url{http://developer.gnome.org/doc/API/2.0/pango/pango-ATSUI-Fonts.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
