\alias{pango-Version-Checking}
\name{pango-Version-Checking}
\title{Version Checking}
\description{Tools for checking Pango version at compile- and run-time.}
\section{Methods and Functions}{
\code{\link{pangoVersion}()}\cr
\code{\link{pangoVersionString}()}\cr
\code{\link{pangoVersionCheck}(required.major, required.minor, required.micro)}\cr
}
\section{Detailed Description}{The capital-letter functions defined here can be used to check the version of
Pango at compile-time, and to \dfn{encode} Pango versions into
integers.
The functions can be used to check the version of the linked Pango library
at run-time.}
\references{\url{https://docs.gtk.org/Pango/}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
