\alias{pangoAttrBackgroundNew}
\name{pangoAttrBackgroundNew}
\title{pangoAttrBackgroundNew}
\description{Create a new background color attribute.}
\usage{pangoAttrBackgroundNew(red, green, blue)}
\arguments{
\item{\code{red}}{[integer]  the red value (ranging from 0 to 65535)}
\item{\code{green}}{[integer]  the green value}
\item{\code{blue}}{[integer]  the blue value}
}
\value{[\code{\link{PangoAttribute}}]  the newly allocated \code{\link{PangoAttribute}}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
