\alias{pangoAttrGravityHintNew}
\name{pangoAttrGravityHintNew}
\title{pangoAttrGravityHintNew}
\description{Create a new gravity hint attribute.}
\usage{pangoAttrGravityHintNew(hint)}
\arguments{\item{\verb{hint}}{[\code{\link{PangoGravityHint}}]  the gravity hint value.}}
\details{  Since 1.16}
\value{[\code{\link{PangoAttribute}}]  the newly allocated \code{\link{PangoAttribute}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
