\alias{pangoAttrGravityNew}
\name{pangoAttrGravityNew}
\title{pangoAttrGravityNew}
\description{Create a new gravity attribute.}
\usage{pangoAttrGravityNew(gravity)}
\arguments{\item{\verb{gravity}}{[\code{\link{PangoGravity}}]  the gravity value; should not be \code{PANGO_GRAVITY_AUTO}.}}
\details{  Since 1.16}
\value{[\code{\link{PangoAttribute}}]  the newly allocated \code{\link{PangoAttribute}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
