\alias{pangoAttributeDestroy}
\name{pangoAttributeDestroy}
\title{pangoAttributeDestroy}
\description{Destroy a \code{\link{PangoAttribute}} and free all associated memory.}
\usage{pangoAttributeDestroy(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoAttribute}}]  a \code{\link{PangoAttribute}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
