\alias{pangoAttributeEqual}
\name{pangoAttributeEqual}
\title{pangoAttributeEqual}
\description{Compare two attributes for equality. This compares only the
actual value of the two attributes and not the ranges that the
attributes apply to.}
\usage{pangoAttributeEqual(object, attr2)}
\arguments{
\item{\verb{object}}{[\code{\link{PangoAttribute}}]  a \code{\link{PangoAttribute}}}
\item{\verb{attr2}}{[\code{\link{PangoAttribute}}]  another \code{\link{PangoAttribute}}}
}
\value{[logical] \code{TRUE} if the two attributes have the same value.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
