\alias{pangoAttributeInit}
\name{pangoAttributeInit}
\title{pangoAttributeInit}
\description{Initializes \code{attr}'s klass to \code{klass}, it's start_index to zero,
and end_index to \code{G_MAXUINT} such that the attribute applies
to the entire text by default.}
\usage{pangoAttributeInit(attr, klass)}
\arguments{
\item{\verb{attr}}{[\verb{\link{PangoAttribute}}]  a \verb{\link{PangoAttribute}}}
\item{\verb{klass}}{[\verb{\link{PangoAttrClass}}]  a \verb{PangoAttributeClass}}
}
\details{  Since  1.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
