\alias{pangoAttrIteratorCopy}
\name{pangoAttrIteratorCopy}
\title{pangoAttrIteratorCopy}
\description{Copy a \verb{\link{PangoAttrIterator}}}
\usage{pangoAttrIteratorCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoAttrIterator}}]  a \verb{\link{PangoAttrIterator}}.}}
\value{[\verb{\link{PangoAttrIterator}}]  the newly allocated \verb{\link{PangoAttrIterator}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
