\alias{pangoAttrIteratorDestroy}
\name{pangoAttrIteratorDestroy}
\title{pangoAttrIteratorDestroy}
\description{Destroy a \code{\link{PangoAttrIterator}} and free all associated memory.}
\usage{pangoAttrIteratorDestroy(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoAttrIterator}}]  a \code{\link{PangoAttrIterator}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
