\alias{pangoAttrIteratorNext}
\name{pangoAttrIteratorNext}
\title{pangoAttrIteratorNext}
\description{Advance the iterator until the next change of style.}
\usage{pangoAttrIteratorNext(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoAttrIterator}}]  a \code{\link{PangoAttrIterator}}}}
\value{[logical]  \code{FALSE} if the iterator is at the end of the list, otherwise \code{TRUE}}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
