\alias{pangoAttrLetterSpacingNew}
\name{pangoAttrLetterSpacingNew}
\title{pangoAttrLetterSpacingNew}
\description{Create a new letter-spacing attribute.}
\usage{pangoAttrLetterSpacingNew(letter.spacing)}
\arguments{\item{\verb{letter.spacing}}{[integer]  amount of extra space to add between graphemes
  of the text, in Pango units.}}
\details{  Since 1.6}
\value{[\code{\link{PangoAttribute}}]  the newly allocated \code{\link{PangoAttribute}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
