\alias{pangoAttrListGetIterator}
\name{pangoAttrListGetIterator}
\title{pangoAttrListGetIterator}
\description{Create a iterator initialized to the beginning of the list.}
\usage{pangoAttrListGetIterator(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoAttrList}}]  a \verb{\link{PangoAttrList}}}}
\value{[\verb{\link{PangoAttrIterator}}]  the newly allocated \verb{\link{PangoAttrIterator}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
