\alias{pangoAttrScaleNew}
\name{pangoAttrScaleNew}
\title{pangoAttrScaleNew}
\description{Create a new font size scale attribute. The base font for the
affected text will have its size multiplied by \code{scale.factor}.}
\usage{pangoAttrScaleNew(scale.factor)}
\arguments{\item{\code{scale.factor}}{[numeric]  factor to scale the font}}
\value{[\code{\link{PangoAttribute}}]  the newly allocated \code{\link{PangoAttribute}}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
