\alias{pangoAttrStrikethroughColorNew}
\name{pangoAttrStrikethroughColorNew}
\title{pangoAttrStrikethroughColorNew}
\description{Create a new strikethrough color attribute. This attribute
modifies the color of strikethrough lines. If not set, strikethrough
lines will use the foreground color.}
\usage{pangoAttrStrikethroughColorNew(red, green, blue)}
\arguments{
\item{\code{red}}{[integer]  the red value (ranging from 0 to 65535)}
\item{\code{green}}{[integer]  the green value}
\item{\code{blue}}{[integer]  the blue value}
}
\value{[\code{\link{PangoAttribute}}]  the new \code{\link{PangoAttribute}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
