\alias{pangoAttrUnderlineColorNew}
\name{pangoAttrUnderlineColorNew}
\title{pangoAttrUnderlineColorNew}
\description{Create a new underline color attribute. This attribute
modifies the color of underlines. If not set, underlines
will use the foreground color.}
\usage{pangoAttrUnderlineColorNew(red, green, blue)}
\arguments{
\item{\code{red}}{[integer]  the red value (ranging from 0 to 65535)}
\item{\code{green}}{[integer]  the green value}
\item{\code{blue}}{[integer]  the blue value}
}
\details{  Since  1.8}
\value{[\code{\link{PangoAttribute}}]  the newly allocated \code{\link{PangoAttribute}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
