\alias{pangoAttrUnderlineNew}
\name{pangoAttrUnderlineNew}
\title{pangoAttrUnderlineNew}
\description{Create a new underline-style object.}
\usage{pangoAttrUnderlineNew(underline)}
\arguments{\item{\code{underline}}{[\code{\link{PangoUnderline}}]  the underline style.}}
\value{[\code{\link{PangoAttribute}}]  the new \code{\link{PangoAttribute}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
