\alias{pangoAttrVariantNew}
\name{pangoAttrVariantNew}
\title{pangoAttrVariantNew}
\description{Create a new font variant attribute (normal or small caps)}
\usage{pangoAttrVariantNew(variant)}
\arguments{\item{\verb{variant}}{[\code{\link{PangoVariant}}]  the variant}}
\value{[\verb{\link{PangoAttribute}}]  the newly allocated \verb{\link{PangoAttribute}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
