\alias{pangoCairoFontGetScaledFont}
\name{pangoCairoFontGetScaledFont}
\title{pangoCairoFontGetScaledFont}
\description{Gets the \verb{\link{CairoScaledFont}} used by \code{font}.}
\usage{pangoCairoFontGetScaledFont(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoCairoFont}}]  a \verb{\link{PangoFont}} from a \verb{\link{PangoCairoFontMap}}}}
\details{  Since  1.18}
\value{[\verb{\link{CairoScaledFont}}]  the \verb{\link{CairoScaledFont}} used by \code{font},
              or \code{NULL} if \code{font} is \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
