\alias{pangoCairoFontMapCreateContext}
\name{pangoCairoFontMapCreateContext}
\title{pangoCairoFontMapCreateContext}
\description{Create a \verb{\link{PangoContext}} for the given fontmap.}
\usage{pangoCairoFontMapCreateContext(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoCairoFontMap}}]  a \verb{\link{PangoCairoFontMap}}}}
\details{  Since  1.10}
\value{[\verb{\link{PangoContext}}] .}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
