\alias{pangoCairoFontMapGetFontType}
\name{pangoCairoFontMapGetFontType}
\title{pangoCairoFontMapGetFontType}
\description{Gets the type of Cairo font backend that \code{fontmap} uses.}
\usage{pangoCairoFontMapGetFontType(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoCairoFontMap}}]  a \verb{\link{PangoCairoFontMap}}}}
\details{  Since  1.18}
\value{[\verb{\link{CairoFontType}}]  the \verb{\link{CairoFontType}} cairo font backend type}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
