\alias{pangoCairoShowLayoutLine}
\name{pangoCairoShowLayoutLine}
\title{pangoCairoShowLayoutLine}
\description{Draws a \code{\link{PangoLayoutLine}} in the specified cairo context.
The origin of the glyphs (the left edge of the line) will
be drawn at the current point of the cairo context.}
\usage{pangoCairoShowLayoutLine(cr, line)}
\arguments{
\item{\verb{cr}}{a Cairo context}
\item{\verb{line}}{a \code{\link{PangoLayoutLine}}}
}
\details{Since 1.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
