\alias{pangoColorCopy}
\name{pangoColorCopy}
\title{pangoColorCopy}
\description{ Primarily used by language bindings,
not that useful otherwise (since colors can just be copied
by assignment in C).}
\usage{pangoColorCopy(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoColor}}]  color to copy}}
\value{[\code{\link{PangoColor}}]  an allocated \code{\link{PangoColor}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
