\alias{pangoColorToString}
\name{pangoColorToString}
\title{pangoColorToString}
\description{Returns a textual specification of \code{color} in the hexadecimal form
\code{#rrrrggggbbbb}, where \code{r},
\code{g} and \code{b} are hex digits representing
the red, green, and blue components respectively.}
\usage{pangoColorToString(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoColor}}]  a \code{\link{PangoColor}}}}
\details{  Since 1.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
