\alias{pangoContextGetBaseGravity}
\name{pangoContextGetBaseGravity}
\title{pangoContextGetBaseGravity}
\description{Retrieves the base gravity for the context. See
\code{\link{pangoContextSetBaseGravity}}.}
\usage{pangoContextGetBaseGravity(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoContext}}]  a \code{\link{PangoContext}}}}
\details{  Since 1.16}
\value{[\code{\link{PangoGravity}}]  the base gravity for the context.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
