\alias{pangoContextGetFontMap}
\name{pangoContextGetFontMap}
\title{pangoContextGetFontMap}
\description{Gets the \verb{PangoFontmap} used to look up fonts for this context.}
\usage{pangoContextGetFontMap(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoContext}}]  a \verb{\link{PangoContext}}}}
\details{  Since  1.6}
\value{[\verb{\link{PangoFontMap}}]  the font map for the \verb{\link{PangoContext}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
