\alias{pangoContextGetGravityHint}
\name{pangoContextGetGravityHint}
\title{pangoContextGetGravityHint}
\description{Retrieves the gravity hint for the context. See
\code{\link{pangoContextSetGravityHint}} for details.}
\usage{pangoContextGetGravityHint(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoContext}}]  a \code{\link{PangoContext}}}}
\details{  Since 1.16}
\value{[\code{\link{PangoGravityHint}}]  the gravity hint for the context.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
