\alias{pangoContextSetBaseGravity}
\name{pangoContextSetBaseGravity}
\title{pangoContextSetBaseGravity}
\description{Sets the base gravity for the context.}
\usage{pangoContextSetBaseGravity(object, gravity)}
\arguments{
\item{\verb{object}}{[\code{\link{PangoContext}}]  a \code{\link{PangoContext}}}
\item{\verb{gravity}}{[\code{\link{PangoGravity}}]  the new base gravity}
}
\details{The base gravity is used in laying vertical text out.  
  Since 1.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
