\alias{pangoContextSetFontDescription}
\name{pangoContextSetFontDescription}
\title{pangoContextSetFontDescription}
\description{Set the default font description for the context}
\usage{pangoContextSetFontDescription(object, desc)}
\arguments{
\item{\verb{object}}{a \code{\link{PangoContext}}}
\item{\verb{desc}}{the new pango font description}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
