\alias{pangoCoverageGet}
\name{pangoCoverageGet}
\title{pangoCoverageGet}
\description{Determine whether a particular index is covered by \code{coverage}}
\usage{pangoCoverageGet(object, index)}
\arguments{
\item{\code{object}}{[\code{\link{PangoCoverage}}]  a \code{\link{PangoCoverage}}}
\item{\code{index}}{[integer]  the index to check}
}
\value{[\code{\link{PangoCoverageLevel}}]  }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
