\alias{pangoCoverageNew}
\name{pangoCoverageNew}
\title{pangoCoverageNew}
\description{Create a new \code{\link{PangoCoverage}}}
\usage{pangoCoverageNew()}
\value{[\code{\link{PangoCoverage}}]  the newly allocated \code{\link{PangoCoverage}},
              initialized to \code{PANGO_COVERAGE_NONE}
              with a reference count of one.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
