\alias{pangoCoverageSet}
\name{pangoCoverageSet}
\title{pangoCoverageSet}
\description{Modify a particular index within \code{coverage}}
\usage{pangoCoverageSet(object, index, level)}
\arguments{
\item{\code{object}}{[\code{\link{PangoCoverage}}]  a \code{\link{PangoCoverage}}}
\item{\code{index}}{[integer]  the index to modify}
\item{\code{level}}{[\code{\link{PangoCoverageLevel}}]  the new level for \code{index.}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
