\alias{pangoFontDescribe}
\name{pangoFontDescribe}
\title{pangoFontDescribe}
\description{Returns a description of the font.}
\usage{pangoFontDescribe(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoFont}}]  a \code{\link{PangoFont}}}}
\value{[\code{\link{PangoFontDescription}}]  a newly-allocated \code{\link{PangoFontDescription}} object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
