\alias{pangoFontDescribeWithAbsoluteSize}
\name{pangoFontDescribeWithAbsoluteSize}
\title{pangoFontDescribeWithAbsoluteSize}
\description{Returns a description of the font, with absolute font size set
(in device units). Use \code{\link{pangoFontDescribe}} if you want the font
size in points.}
\usage{pangoFontDescribeWithAbsoluteSize(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoFont}}]  a \verb{\link{PangoFont}}}}
\details{  Since  1.14}
\value{[\verb{\link{PangoFontDescription}}]  a newly-allocated \verb{\link{PangoFontDescription}} object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
