\alias{pangoFontDescriptionGetGravity}
\name{pangoFontDescriptionGetGravity}
\title{pangoFontDescriptionGetGravity}
\description{Gets the gravity field of a font description. See
\code{\link{pangoFontDescriptionSetGravity}}.}
\usage{pangoFontDescriptionGetGravity(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoFontDescription}}]  a \code{\link{PangoFontDescription}}}}
\details{  Since 1.16}
\value{[\code{\link{PangoGravity}}]  the gravity field for the font description. Use
  \code{\link{pangoFontDescriptionGetSetFields}} to find out if
  the field was explicitly set or not.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
