\alias{pangoFontDescriptionGetSetFields}
\name{pangoFontDescriptionGetSetFields}
\title{pangoFontDescriptionGetSetFields}
\description{Determines which fields in a font description have been set.}
\usage{pangoFontDescriptionGetSetFields(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoFontDescription}}]  a \verb{\link{PangoFontDescription}}}}
\value{[\code{\link{PangoFontMask}}]  a bitmask with bits set corresponding to the
  fields in \code{desc} that have been set.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
