\alias{pangoFontDescriptionGetStretch}
\name{pangoFontDescriptionGetStretch}
\title{pangoFontDescriptionGetStretch}
\description{Gets the stretch field of a font description.
See \code{\link{pangoFontDescriptionSetStretch}}.}
\usage{pangoFontDescriptionGetStretch(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoFontDescription}}]  a \code{\link{PangoFontDescription}}.}}
\value{[\code{\link{PangoStretch}}]  the stretch field for the font description. Use
  \code{\link{pangoFontDescriptionGetSetFields}} to find out if
  the field was explicitly set or not.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
