\alias{pangoFontDescriptionGetStyle}
\name{pangoFontDescriptionGetStyle}
\title{pangoFontDescriptionGetStyle}
\description{Gets the style field of a \code{\link{PangoFontDescription}}. See
\code{\link{pangoFontDescriptionSetStyle}}.}
\usage{pangoFontDescriptionGetStyle(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoFontDescription}}]  a \code{\link{PangoFontDescription}}}}
\value{[\code{\link{PangoStyle}}]  the style field for the font description.
  Use \code{\link{pangoFontDescriptionGetSetFields}} to find out if
  the field was explicitly set or not.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
