\alias{pangoFontDescriptionGetVariant}
\name{pangoFontDescriptionGetVariant}
\title{pangoFontDescriptionGetVariant}
\description{Gets the variant field of a \verb{\link{PangoFontDescription}}. See
\code{\link{pangoFontDescriptionSetVariant}}.}
\usage{pangoFontDescriptionGetVariant(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoFontDescription}}]  a \verb{\link{PangoFontDescription}}.}}
\value{[\code{\link{PangoVariant}}]  the variant field for the font description. Use
  \code{\link{pangoFontDescriptionGetSetFields}} to find out if
  the field was explicitly set or not.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
