\alias{pangoFontDescriptionGetWeight}
\name{pangoFontDescriptionGetWeight}
\title{pangoFontDescriptionGetWeight}
\description{Gets the weight field of a font description. See
\code{\link{pangoFontDescriptionSetWeight}}.}
\usage{pangoFontDescriptionGetWeight(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoFontDescription}}]  a \code{\link{PangoFontDescription}}}}
\value{[\code{\link{PangoWeight}}]  the weight field for the font description. Use
  \code{\link{pangoFontDescriptionGetSetFields}} to find out if
  the field was explicitly set or not.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
