\alias{pangoFontDescriptionSetStretch}
\name{pangoFontDescriptionSetStretch}
\title{pangoFontDescriptionSetStretch}
\description{Sets the stretch field of a font description. The stretch field
specifies how narrow or wide the font should be.}
\usage{pangoFontDescriptionSetStretch(object, stretch)}
\arguments{
\item{\verb{object}}{a \code{\link{PangoFontDescription}}}
\item{\verb{stretch}}{the stretch for the font description}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
