\alias{pangoFontDescriptionSetVariant}
\name{pangoFontDescriptionSetVariant}
\title{pangoFontDescriptionSetVariant}
\description{Sets the variant field of a font description. The \code{\link{PangoVariant}}
can either be \code{PANGO_VARIANT_NORMAL} or \code{PANGO_VARIANT_SMALL_CAPS}.}
\usage{pangoFontDescriptionSetVariant(object, variant)}
\arguments{
\item{\verb{object}}{[\code{\link{PangoFontDescription}}]  a \code{\link{PangoFontDescription}}}
\item{\verb{variant}}{[\code{\link{PangoVariant}}]  the variant type for the font description.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
