\alias{pangoFontDescriptionSetWeight}
\name{pangoFontDescriptionSetWeight}
\title{pangoFontDescriptionSetWeight}
\description{Sets the weight field of a font description. The weight field
specifies how bold or light the font should be. In addition
to the values of the \code{\link{PangoWeight}} enumeration, other intermediate
numeric values are possible.}
\usage{pangoFontDescriptionSetWeight(object, weight)}
\arguments{
\item{\verb{object}}{[\code{\link{PangoFontDescription}}]  a \code{\link{PangoFontDescription}}}
\item{\verb{weight}}{[\code{\link{PangoWeight}}]  the weight for the font description.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
