\alias{pangoFontDescriptionToString}
\name{pangoFontDescriptionToString}
\title{pangoFontDescriptionToString}
\description{Creates a string representation of a font description. See
\code{\link{pangoFontDescriptionFromString}} for a description of the
format of the string representation. The family list in the
string description will only have a terminating comma if the
last word of the list is a valid style option.}
\usage{pangoFontDescriptionToString(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoFontDescription}}]  a \verb{\link{PangoFontDescription}}}}
\value{[char] .}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
