\alias{pangoFontFaceGetFaceName}
\name{pangoFontFaceGetFaceName}
\title{pangoFontFaceGetFaceName}
\description{Gets a name representing the style of this face among the
different faces in the \code{\link{PangoFontFamily}} for the face. This
name is unique among all faces in the family and is suitable
for displaying to users.}
\usage{pangoFontFaceGetFaceName(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoFontFace}}]  a \code{\link{PangoFontFace}}.}}
\value{[char]  the face name for the face.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
