\alias{pangoFontGetCoverage}
\name{pangoFontGetCoverage}
\title{pangoFontGetCoverage}
\description{Computes the coverage map for a given font and language tag.}
\usage{pangoFontGetCoverage(object, language)}
\arguments{
\item{\verb{object}}{[\verb{\link{PangoFont}}]  a \verb{\link{PangoFont}}}
\item{\verb{language}}{[\verb{\link{PangoLanguage}}]  the language tag}
}
\value{[\verb{\link{PangoCoverage}}]  a newly-allocated \verb{\link{PangoCoverage}} object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
