\alias{pangoFontGetFontMap}
\name{pangoFontGetFontMap}
\title{pangoFontGetFontMap}
\description{Gets the font map for which the font was created.}
\usage{pangoFontGetFontMap(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoFont}}]  a \verb{\link{PangoFont}}}}
\details{  Since  1.10}
\value{[\verb{\link{PangoFontMap}}]  the \verb{\link{PangoFontMap}} for the font, or \code{NULL} if \code{font} is \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
