\alias{pangoFontMetricsGetStrikethroughPosition}
\name{pangoFontMetricsGetStrikethroughPosition}
\title{pangoFontMetricsGetStrikethroughPosition}
\description{Gets the suggested position to draw the strikethrough.
The value returned is the distance \emph{above} the
baseline of the top of the strikethrough.}
\usage{pangoFontMetricsGetStrikethroughPosition(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoFontMetrics}}]  a \code{\link{PangoFontMetrics}} structure}}
\details{  Since 1.6}
\value{[integer]  the suggested strikethrough position, in Pango units.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
