\alias{pangoFontMetricsGetStrikethroughThickness}
\name{pangoFontMetricsGetStrikethroughThickness}
\title{pangoFontMetricsGetStrikethroughThickness}
\description{Gets the suggested thickness to draw for the strikethrough.}
\usage{pangoFontMetricsGetStrikethroughThickness(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoFontMetrics}}]  a \code{\link{PangoFontMetrics}} structure}}
\details{  Since 1.6}
\value{[integer]  the suggested strikethrough thickness, in Pango units.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
