\alias{pangoFontMetricsGetUnderlineThickness}
\name{pangoFontMetricsGetUnderlineThickness}
\title{pangoFontMetricsGetUnderlineThickness}
\description{Gets the suggested thickness to draw for the underline.}
\usage{pangoFontMetricsGetUnderlineThickness(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoFontMetrics}}]  a \code{\link{PangoFontMetrics}} structure}}
\details{  Since 1.6}
\value{[integer]  the suggested underline thickness, in Pango units.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
