\alias{pangoGlyphStringCopy}
\name{pangoGlyphStringCopy}
\title{pangoGlyphStringCopy}
\description{Copy a glyph string and associated storage.}
\usage{pangoGlyphStringCopy(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoGlyphString}}]  a \code{\link{PangoGlyphString}}.}}
\value{[\code{\link{PangoGlyphString}}]  the newly allocated \code{\link{PangoGlyphString}}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
