\alias{pangoGravityToRotation}
\name{pangoGravityToRotation}
\title{pangoGravityToRotation}
\description{Converts a \code{\link{PangoGravity}} value to its rotation value.
\code{gravity} should not be \code{PANGO_GRAVITY_AUTO}.}
\usage{pangoGravityToRotation(base.gravity)}
\arguments{\item{\code{base.gravity}}{[\code{\link{PangoGravity}}]  gravity to query}}
\details{  Since  1.16}
\value{[numeric]  the rotation value corresponding to \code{gravity}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
