\alias{pangoItemCopy}
\name{pangoItemCopy}
\title{pangoItemCopy}
\description{Copy an existing \verb{\link{PangoItem}} structure.}
\usage{pangoItemCopy(item)}
\arguments{\item{\verb{item}}{[\verb{\link{PangoItem}}]  a \verb{\link{PangoItem}}, may be \code{NULL}}}
\value{[\verb{\link{PangoItem}}]  the newly allocated \verb{\link{PangoItem}}, or \code{NULL} if
              \code{item} was NULL.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
