\alias{pangoItemNew}
\name{pangoItemNew}
\title{pangoItemNew}
\description{Creates a new \code{\link{PangoItem}} structure initialized to default values.}
\usage{pangoItemNew()}
\value{[\code{\link{PangoItem}}]  the newly allocated \code{\link{PangoItem}}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
