\alias{pangoLanguageToString}
\name{pangoLanguageToString}
\title{pangoLanguageToString}
\description{Get the string representation of a \code{\link{PangoLanguage}}.}
\usage{
pangoLanguageToString(object)
}
\arguments{
\item{object}{The \code{\link{PangoLanguage}} to convert to a string.}
}
\value{
a string representation (the name) of the language
}
\author{Michael Lawrence}
\keyword{internal}
\keyword{interface}
