\alias{pangoLayoutCopy}
\name{pangoLayoutCopy}
\title{pangoLayoutCopy}
\description{Does a deep copy-by-value of the \code{src} layout. The attribute list,
tab list, and text from the original layout are all copied by
value.}
\usage{pangoLayoutCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoLayout}}]  a \verb{\link{PangoLayout}}}}
\value{[\verb{\link{PangoLayout}}]  the newly allocated \verb{\link{PangoLayout}}, with a reference
              count of one.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
