\alias{pangoLayoutGetAlignment}
\name{pangoLayoutGetAlignment}
\title{pangoLayoutGetAlignment}
\description{Gets the alignment for the layout: how partial lines are
positioned within the horizontal space available.}
\usage{pangoLayoutGetAlignment(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoLayout}}]  a \verb{\link{PangoLayout}}}}
\value{[\code{\link{PangoAlignment}}]  the alignment.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
