\alias{pangoLayoutGetAttributes}
\name{pangoLayoutGetAttributes}
\title{pangoLayoutGetAttributes}
\description{Gets the attribute list for the layout, if any.}
\usage{pangoLayoutGetAttributes(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}}}
\value{[\code{\link{PangoAttrList}}]  a \code{\link{PangoAttrList}}.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
