\alias{pangoLayoutGetAutoDir}
\name{pangoLayoutGetAutoDir}
\title{pangoLayoutGetAutoDir}
\description{Gets whether to calculate the bidirectional base direction
for the layout according to the contents of the layout.
See \code{\link{pangoLayoutSetAutoDir}}.}
\usage{pangoLayoutGetAutoDir(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}}}
\value{[logical]  if \code{TRUE}, the bidirectional base direction
  is computed from the layout's contents.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
