\alias{pangoLayoutGetBaseline}
\name{pangoLayoutGetBaseline}
\title{pangoLayoutGetBaseline}
\description{Gets the Y position of baseline of the first line in \code{layout}.}
\usage{pangoLayoutGetBaseline(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}}}
\details{  Since 1.22}
\value{[integer]  baseline of first line, from top of \code{layout}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
