\alias{pangoLayoutGetHeight}
\name{pangoLayoutGetHeight}
\title{pangoLayoutGetHeight}
\description{Gets the height of layout used for ellipsization.  See
\code{\link{pangoLayoutSetHeight}} for details.}
\usage{pangoLayoutGetHeight(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}}}
\details{  Since 1.20}
\value{[integer]  the height, in Pango units if positive, or
number of lines if negative.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
