\alias{pangoLayoutGetIndent}
\name{pangoLayoutGetIndent}
\title{pangoLayoutGetIndent}
\description{Gets the paragraph indent width in pango units. A negative value
indicates a hanging indent.}
\usage{pangoLayoutGetIndent(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}}}
\value{[integer]  the indent}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
