\alias{pangoLayoutGetIter}
\name{pangoLayoutGetIter}
\title{pangoLayoutGetIter}
\description{Returns an iterator to iterate over the visual extents of the layout.}
\usage{pangoLayoutGetIter(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}}}
\value{[\code{\link{PangoLayoutIter}}] .}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
