\alias{pangoLayoutGetLineCount}
\name{pangoLayoutGetLineCount}
\title{pangoLayoutGetLineCount}
\description{Retrieves the count of lines for the \code{layout}.}
\usage{pangoLayoutGetLineCount(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoLayout}}]  \code{\link{PangoLayout}}}}
\value{[integer]  the line count.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
