\alias{pangoLayoutGetSingleParagraphMode}
\name{pangoLayoutGetSingleParagraphMode}
\title{pangoLayoutGetSingleParagraphMode}
\description{Obtains the value set by \code{\link{pangoLayoutSetSingleParagraphMode}}.}
\usage{pangoLayoutGetSingleParagraphMode(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoLayout}}]  a \verb{\link{PangoLayout}}}}
\value{[logical] \code{TRUE} if the layout does not break paragraphs at
paragraph separator characters, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
