\alias{pangoLayoutGetSpacing}
\name{pangoLayoutGetSpacing}
\title{pangoLayoutGetSpacing}
\description{Gets the amount of spacing in \code{integer} between the lines of the
layout.}
\usage{pangoLayoutGetSpacing(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}}}
\value{[integer]  the spacing.}
\section{encoding}{UTF-8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
