\alias{pangoLayoutGetText}
\name{pangoLayoutGetText}
\title{pangoLayoutGetText}
\description{Gets the text in the layout.}
\usage{pangoLayoutGetText(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoLayout}}]  a \verb{\link{PangoLayout}}}}
\value{[char]  the text in the \code{layout}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
