\alias{pangoLayoutGetWidth}
\name{pangoLayoutGetWidth}
\title{pangoLayoutGetWidth}
\description{Gets the width to which the lines of the \code{\link{PangoLayout}} should be wrapped.}
\usage{pangoLayoutGetWidth(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}}}
\value{[integer]  the width}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
