\alias{pangoLayoutGetWrap}
\name{pangoLayoutGetWrap}
\title{pangoLayoutGetWrap}
\description{Gets the wrap mode for the layout.}
\usage{pangoLayoutGetWrap(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}}}
\value{[\code{\link{PangoWrapMode}}]  Active wrap mode.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
