\alias{pangoLayoutIterAtLastLine}
\name{pangoLayoutIterAtLastLine}
\title{pangoLayoutIterAtLastLine}
\description{Determines whether \code{iter} is on the last line of the layout.}
\usage{pangoLayoutIterAtLastLine(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoLayoutIter}}]  a \code{\link{PangoLayoutIter}}}}
\value{[logical] \code{TRUE} if \code{iter} is on the last line.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
