\alias{pangoLayoutIterGetLayout}
\name{pangoLayoutIterGetLayout}
\title{pangoLayoutIterGetLayout}
\description{Gets the layout associated with a \verb{\link{PangoLayoutIter}}.}
\usage{pangoLayoutIterGetLayout(iter)}
\arguments{\item{\verb{iter}}{[\verb{\link{PangoLayoutIter}}]  a \verb{\link{PangoLayoutIter}}}}
\details{  Since  1.20}
\value{[\verb{\link{PangoLayout}}]  the layout associated with \code{iter}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
