\alias{pangoLayoutIterGetLine}
\name{pangoLayoutIterGetLine}
\title{pangoLayoutIterGetLine}
\description{Gets the current line.}
\usage{pangoLayoutIterGetLine(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoLayoutIter}}]  a \code{\link{PangoLayoutIter}}}}
\value{[\code{\link{PangoLayoutLine}}]  the current line}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
