\alias{pangoLayoutIterNextChar}
\name{pangoLayoutIterNextChar}
\title{pangoLayoutIterNextChar}
\description{Moves \code{iter} forward to the next character in visual order. If \code{iter} was already at
the end of the layout, returns \code{FALSE}.}
\usage{pangoLayoutIterNextChar(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoLayoutIter}}]  a \code{\link{PangoLayoutIter}}}}
\value{[logical]  whether motion was possible.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
