\alias{pangoLayoutIterNextCluster}
\name{pangoLayoutIterNextCluster}
\title{pangoLayoutIterNextCluster}
\description{Moves \code{iter} forward to the next cluster in visual order. If \code{iter}
was already at the end of the layout, returns \code{FALSE}.}
\usage{pangoLayoutIterNextCluster(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoLayoutIter}}]  a \code{\link{PangoLayoutIter}}}}
\value{[logical]  whether motion was possible.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
