\alias{pangoLayoutIterNextLine}
\name{pangoLayoutIterNextLine}
\title{pangoLayoutIterNextLine}
\description{Moves \code{iter} forward to the start of the next line. If \code{iter} is
already on the last line, returns \code{FALSE}.}
\usage{pangoLayoutIterNextLine(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoLayoutIter}}]  a \code{\link{PangoLayoutIter}}}}
\value{[logical]  whether motion was possible.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
