\alias{pangoLayoutIterNextRun}
\name{pangoLayoutIterNextRun}
\title{pangoLayoutIterNextRun}
\description{Moves \code{iter} forward to the next run in visual order. If \code{iter} was
already at the end of the layout, returns \code{FALSE}.}
\usage{pangoLayoutIterNextRun(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoLayoutIter}}]  a \code{\link{PangoLayoutIter}}}}
\value{[logical]  whether motion was possible.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
