\alias{pangoLayoutNew}
\name{pangoLayoutNew}
\title{pangoLayoutNew}
\description{Create a new \verb{\link{PangoLayout}} object with attributes initialized to
default values for a particular \verb{\link{PangoContext}}.}
\usage{pangoLayoutNew(context)}
\arguments{\item{\verb{context}}{[\verb{\link{PangoContext}}]  a \verb{\link{PangoContext}}}}
\value{[\verb{\link{PangoLayout}}]  the newly allocated \verb{\link{PangoLayout}}, with a reference
              count of one.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
