\alias{pangoLayoutSetAlignment}
\name{pangoLayoutSetAlignment}
\title{pangoLayoutSetAlignment}
\description{Sets the alignment for the layout: how partial lines are
positioned within the horizontal space available.}
\usage{pangoLayoutSetAlignment(object, alignment)}
\arguments{
\item{\verb{object}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}}
\item{\verb{alignment}}{[\code{\link{PangoAlignment}}]  the alignment}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
