\alias{pangoLayoutSetAttributes}
\name{pangoLayoutSetAttributes}
\title{pangoLayoutSetAttributes}
\description{Sets the text attributes for a layout object.}
\usage{pangoLayoutSetAttributes(object, attrs)}
\arguments{
\item{\code{object}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}}
\item{\code{attrs}}{[\code{\link{PangoAttrList}}]  a \code{\link{PangoAttrList}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
