\alias{pangoLayoutSetSpacing}
\name{pangoLayoutSetSpacing}
\title{pangoLayoutSetSpacing}
\description{Sets the amount of spacing between the lines of the layout.}
\usage{pangoLayoutSetSpacing(object, spacing)}
\arguments{
\item{\code{object}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}.}
\item{\code{spacing}}{[integer]  the amount of spacing}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
