\alias{pangoMatrixCopy}
\name{pangoMatrixCopy}
\title{pangoMatrixCopy}
\description{Copies a \verb{\link{PangoMatrix}}.}
\usage{pangoMatrixCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoMatrix}}]  a \verb{\link{PangoMatrix}}, may be \code{NULL}}}
\details{  Since  1.6}
\value{[\verb{\link{PangoMatrix}}]  the newly allocated \verb{\link{PangoMatrix}}, or \code{NULL} if
              \code{matrix} was \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
