\alias{pangoMatrixRotate}
\name{pangoMatrixRotate}
\title{pangoMatrixRotate}
\description{Changes the transformation represented by \code{matrix} to be the
transformation given by first rotating by \code{degrees} degrees
counter-clockwise then applying the original transformation.}
\usage{pangoMatrixRotate(object, degrees)}
\arguments{
\item{\verb{object}}{[\code{\link{PangoMatrix}}]  a \code{\link{PangoMatrix}}}
\item{\verb{degrees}}{[numeric]  degrees to rotate counter-clockwise}
}
\details{  Since 1.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
