\alias{pangoMatrixScale}
\name{pangoMatrixScale}
\title{pangoMatrixScale}
\description{Changes the transformation represented by \code{matrix} to be the
transformation given by first scaling by \code{sx} in the X direction
and \code{sy} in the Y direction then applying the original
transformation.}
\usage{pangoMatrixScale(object, scale.x, scale.y)}
\arguments{
\item{\verb{object}}{[\code{\link{PangoMatrix}}]  a \code{\link{PangoMatrix}}}
\item{\verb{scale.x}}{[numeric]  amount to scale by in X direction}
\item{\verb{scale.y}}{[numeric]  amount to scale by in Y direction}
}
\details{  Since 1.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
