\alias{pangoMatrixTransformPoint}
\name{pangoMatrixTransformPoint}
\title{pangoMatrixTransformPoint}
\description{Transforms the point (\code{x}, \code{y}) by \code{matrix}.}
\usage{pangoMatrixTransformPoint(object, x, y)}
\arguments{
\item{\code{object}}{[\code{\link{PangoMatrix}}]  a \code{\link{PangoMatrix}}, or \code{NULL}}
\item{\code{x}}{[numeric]  in/out X position}
\item{\code{y}}{[numeric]  in/out Y position}
}
\details{  Since  1.16}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
