\alias{pangoRendererDeactivate}
\name{pangoRendererDeactivate}
\title{pangoRendererDeactivate}
\description{Cleans up after rendering operations on \code{renderer}. See
docs for \code{\link{pangoRendererActivate}}.}
\usage{pangoRendererDeactivate(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoRenderer}}]  a \verb{\link{PangoRenderer}}}}
\details{  Since  1.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
