\alias{pangoRendererDrawGlyph}
\name{pangoRendererDrawGlyph}
\title{pangoRendererDrawGlyph}
\description{Draws a single glyph with coordinates in device space.}
\usage{pangoRendererDrawGlyph(object, font, glyph, x, y)}
\arguments{
\item{\verb{object}}{[\verb{\link{PangoRenderer}}]  a \verb{\link{PangoRenderer}}}
\item{\verb{font}}{[\verb{\link{PangoFont}}]  a \verb{\link{PangoFont}}}
\item{\verb{glyph}}{[numeric]  the glyph index of a single glyph}
\item{\verb{x}}{[numeric]  X coordinate of left edge of baseline of glyph}
\item{\verb{y}}{[numeric]  Y coordinate of left edge of baseline of glyph}
}
\details{  Since  1.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
