\alias{pangoRendererDrawLayout}
\name{pangoRendererDrawLayout}
\title{pangoRendererDrawLayout}
\description{Draws \code{layout} with the specified \verb{\link{PangoRenderer}}.}
\usage{pangoRendererDrawLayout(object, layout, x, y)}
\arguments{
\item{\verb{object}}{[\verb{\link{PangoRenderer}}]  a \verb{\link{PangoRenderer}}}
\item{\verb{layout}}{[\verb{\link{PangoLayout}}]  a \verb{\link{PangoLayout}}}
\item{\verb{x}}{[integer]  X position of left edge of baseline, in user space coordinates
  in Pango units.}
\item{\verb{y}}{[integer]  Y position of left edge of baseline, in user space coordinates
   in Pango units.}
}
\details{  Since  1.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
