\alias{pangoRendererSetColor}
\name{pangoRendererSetColor}
\title{pangoRendererSetColor}
\description{Sets the color for part of the rendering.}
\usage{pangoRendererSetColor(object, part, color)}
\arguments{
\item{\verb{object}}{a \code{\link{PangoRenderer}}}
\item{\verb{part}}{the part to change the color of}
\item{\verb{color}}{the new color or \code{NULL} to unset the current color}
}
\details{Since 1.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
