\alias{pangoTabArrayCopy}
\name{pangoTabArrayCopy}
\title{pangoTabArrayCopy}
\description{Copies a \verb{\link{PangoTabArray}}}
\usage{pangoTabArrayCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoTabArray}}] \verb{\link{PangoTabArray}} to copy}}
\value{[\verb{\link{PangoTabArray}}]  the newly allocated \verb{\link{PangoTabArray}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
