\alias{pangoTabArrayGetPositionsInPixels}
\name{pangoTabArrayGetPositionsInPixels}
\title{pangoTabArrayGetPositionsInPixels}
\description{Returns TRUE if the tab positions are in pixels, FALSE if they are
in Pango units.}
\usage{pangoTabArrayGetPositionsInPixels(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoTabArray}}]  a \code{\link{PangoTabArray}}}}
\value{[logical]  whether positions are in pixels}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
