\alias{pangoTabArrayGetSize}
\name{pangoTabArrayGetSize}
\title{pangoTabArrayGetSize}
\description{Gets the number of tab stops in \code{tab.array}.}
\usage{pangoTabArrayGetSize(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoTabArray}}]  a \code{\link{PangoTabArray}}}}
\value{[integer]  the number of tab stops in the list.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
