\alias{pangoUnitsFromDouble}
\name{pangoUnitsFromDouble}
\title{pangoUnitsFromDouble}
\description{Converts a floating-point number to Pango units: multiplies
it by \code{PANGO_SCALE} and rounds to nearest integer.}
\usage{pangoUnitsFromDouble(d)}
\arguments{\item{\verb{d}}{[numeric]  double floating-point value}}
\details{  Since 1.16}
\value{[integer]  the value in Pango units.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
