\alias{pangoUnitsToDouble}
\name{pangoUnitsToDouble}
\title{pangoUnitsToDouble}
\description{Converts a number in Pango units to floating-point: divides
it by \code{PANGO_SCALE}.}
\usage{pangoUnitsToDouble(i)}
\arguments{\item{\verb{i}}{[integer]  value in Pango units}}
\details{  Since 1.16}
\value{[numeric]  the double value.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
