\alias{pangoVersionString}
\name{pangoVersionString}
\title{pangoVersionString}
\description{This is similar to the function \code{PANGO_VERSION_STRING} except that
it returns the version of Pango available at run-time, as opposed to
the version available at compile-time.}
\usage{pangoVersionString()}
\details{  Since 1.16}
\value{[char]  A string containing the version of Pango library
  available at run time.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
