\name{stock-items}
\alias{stock-items}
\title{GTK Stock Items}
\alias{GTK_STOCK_ABOUT}
\alias{GTK_STOCK_ADD}
\alias{GTK_STOCK_APPLY}
\alias{GTK_STOCK_BOLD}
\alias{GTK_STOCK_CANCEL}
\alias{GTK_STOCK_CDROM}
\alias{GTK_STOCK_CLEAR}
\alias{GTK_STOCK_CLOSE}
\alias{GTK_STOCK_COLOR_PICKER}
\alias{GTK_STOCK_CONNECT}
\alias{GTK_STOCK_CONVERT}
\alias{GTK_STOCK_COPY}
\alias{GTK_STOCK_CUT}
\alias{GTK_STOCK_DELETE}
\alias{GTK_STOCK_DIALOG_AUTHENTICATION}
\alias{GTK_STOCK_DIALOG_ERROR}
\alias{GTK_STOCK_DIALOG_INFO}
\alias{GTK_STOCK_DIALOG_QUESTION}
\alias{GTK_STOCK_DIALOG_WARNING}
\alias{GTK_STOCK_DIRECTORY}
\alias{GTK_STOCK_DISCONNECT}
\alias{GTK_STOCK_DND}
\alias{GTK_STOCK_DND_MULTIPLE}
\alias{GTK_STOCK_EDIT}
\alias{GTK_STOCK_EXECUTE}
\alias{GTK_STOCK_FILE}
\alias{GTK_STOCK_FIND}
\alias{GTK_STOCK_FIND_AND_REPLACE}
\alias{GTK_STOCK_FLOPPY}
\alias{GTK_STOCK_FULLSCREEN}
\alias{GTK_STOCK_GOTO_BOTTOM}
\alias{GTK_STOCK_GOTO_FIRST}
\alias{GTK_STOCK_GOTO_LAST}
\alias{GTK_STOCK_GOTO_TOP}
\alias{GTK_STOCK_GO_BACK}
\alias{GTK_STOCK_GO_DOWN}
\alias{GTK_STOCK_GO_FORWARD}
\alias{GTK_STOCK_GO_UP}
\alias{GTK_STOCK_HARDDISK}
\alias{GTK_STOCK_HELP}
\alias{GTK_STOCK_HOME}
\alias{GTK_STOCK_INDENT}
\alias{GTK_STOCK_INDEX}
\alias{GTK_STOCK_INFO}
\alias{GTK_STOCK_ITALIC}
\alias{GTK_STOCK_JUMP_TO}
\alias{GTK_STOCK_JUSTIFY_CENTER}
\alias{GTK_STOCK_JUSTIFY_FILL}
\alias{GTK_STOCK_JUSTIFY_LEFT}
\alias{GTK_STOCK_JUSTIFY_RIGHT}
\alias{GTK_STOCK_LEAVE_FULLSCREEN}
\alias{GTK_STOCK_MEDIA_FORWARD}
\alias{GTK_STOCK_MEDIA_NEXT}
\alias{GTK_STOCK_MEDIA_PAUSE}
\alias{GTK_STOCK_MEDIA_PLAY}
\alias{GTK_STOCK_MEDIA_PREVIOUS}
\alias{GTK_STOCK_MEDIA_RECORD}
\alias{GTK_STOCK_MEDIA_REWIND}
\alias{GTK_STOCK_MEDIA_STOP}
\alias{GTK_STOCK_MISSING_IMAGE}
\alias{GTK_STOCK_NETWORK}
\alias{GTK_STOCK_NEW}
\alias{GTK_STOCK_NO}
\alias{GTK_STOCK_OK}
\alias{GTK_STOCK_OPEN}
\alias{GTK_STOCK_PASTE}
\alias{GTK_STOCK_PREFERENCES}
\alias{GTK_STOCK_PRINT}
\alias{GTK_STOCK_PRINT_PREVIEW}
\alias{GTK_STOCK_PROPERTIES}
\alias{GTK_STOCK_QUIT}
\alias{GTK_STOCK_REDO}
\alias{GTK_STOCK_REFRESH}
\alias{GTK_STOCK_REMOVE}
\alias{GTK_STOCK_REVERT_TO_SAVED}
\alias{GTK_STOCK_SAVE}
\alias{GTK_STOCK_SAVE_AS}
\alias{GTK_STOCK_SELECT_COLOR}
\alias{GTK_STOCK_SELECT_FONT}
\alias{GTK_STOCK_SORT_ASCENDING}
\alias{GTK_STOCK_SORT_DESCENDING}
\alias{GTK_STOCK_SPELL_CHECK}
\alias{GTK_STOCK_STOP}
\alias{GTK_STOCK_STRIKETHROUGH}
\alias{GTK_STOCK_UNDELETE}
\alias{GTK_STOCK_UNDERLINE}
\alias{GTK_STOCK_UNDO}
\alias{GTK_STOCK_UNINDENT}
\alias{GTK_STOCK_YES}
\alias{GTK_STOCK_ZOOM_100}
\alias{GTK_STOCK_ZOOM_FIT}
\alias{GTK_STOCK_ZOOM_IN}
\alias{GTK_STOCK_ZOOM_OUT}
\description{Please see \url{https://developer-old.gnome.org/gtk2/stable/gtk2-Stock-Items.html} for details on stock items.}
\author{Michael Lawrence}
\keyword{internal}
