\alias{atkActionGetType}
\alias{atkComponentGetType}
\alias{atkDocumentGetType}
\alias{atkEditableTextGetType}
\alias{atkGObjectAccessibleGetType}
\alias{atkHyperlinkGetType}
\alias{atkHypertextGetType}
\alias{atkImageGetType}
\alias{atkImplementorGetType}
\alias{atkKeyEventStructGetType}
\alias{atkNoOpObjectFactoryGetType}
\alias{atkNoOpObjectGetType}
\alias{atkObjectFactoryGetType}
\alias{atkObjectGetType}
\alias{atkRegistryGetType}
\alias{atkRelationGetType}
\alias{atkRelationSetGetType}
\alias{atkSelectionGetType}
\alias{atkStateSetGetType}
\alias{atkStreamableContentGetType}
\alias{atkTableGetType}
\alias{atkTextGetType}
\alias{atkUtilGetType}
\alias{atkValueGetType}
\alias{gdkColormapGetType}
\alias{gdkCursorGetType}
\alias{gdkDeviceGetType}
\alias{gdkDisplayGetType}
\alias{gdkDisplayManagerGetType}
\alias{gdkDragContextGetType}
\alias{gdkDrawableGetType}
\alias{gdkEventAnyGetType}
\alias{gdkEventGetType}
\alias{gdkFontGetType}
\alias{gdkGCGetType}
\alias{gdkImageGetType}
\alias{gdkKeymapGetType}
\alias{gdkPangoRendererGetType}
\alias{gdkPixbufAnimationGetType}
\alias{gdkPixbufAnimationIterGetType}
\alias{gdkPixbufLoaderGetType}
\alias{gdkPixmapGetType}
\alias{gdkScreenGetType}
\alias{gdkVisualGetType}
\alias{gdkWindowObjectGetType}
\alias{gladeXMLGetType}
\alias{gtkAboutDialogGetType}
\alias{gtkAccelGroupGetType}
\alias{gtkAccelLabelGetType}
\alias{gtkAccelMapGetType}
\alias{gtkAccessibleGetType}
\alias{gtkActionGetType}
\alias{gtkActionGroupGetType}
\alias{gtkAdjustmentGetType}
\alias{gtkAlignmentGetType}
\alias{gtkArrowGetType}
\alias{gtkAspectFrameGetType}
\alias{gtkBinGetType}
\alias{gtkBorderGetType}
\alias{gtkBoxGetType}
\alias{gtkButtonBoxGetType}
\alias{gtkButtonGetType}
\alias{gtkCListGetType}
\alias{gtkCTreeGetType}
\alias{gtkCTreeNodeGetType}
\alias{gtkCalendarGetType}
\alias{gtkCellEditableGetType}
\alias{gtkCellLayoutGetType}
\alias{gtkCellRendererComboGetType}
\alias{gtkCellRendererGetType}
\alias{gtkCellRendererPixbufGetType}
\alias{gtkCellRendererProgressGetType}
\alias{gtkCellRendererTextGetType}
\alias{gtkCellRendererToggleGetType}
\alias{gtkCellViewGetType}
\alias{gtkCheckButtonGetType}
\alias{gtkCheckMenuItemGetType}
\alias{gtkClipboardGetType}
\alias{gtkColorButtonGetType}
\alias{gtkColorSelectionDialogGetType}
\alias{gtkColorSelectionGetType}
\alias{gtkComboBoxEntryGetType}
\alias{gtkComboBoxGetType}
\alias{gtkComboGetType}
\alias{gtkContainerGetType}
\alias{gtkCurveGetType}
\alias{gtkDialogGetType}
\alias{gtkDrawingAreaGetType}
\alias{gtkEditableGetType}
\alias{gtkEntryCompletionGetType}
\alias{gtkEntryGetType}
\alias{gtkEventBoxGetType}
\alias{gtkExpanderGetType}
\alias{gtkFileChooserButtonGetType}
\alias{gtkFileChooserDialogGetType}
\alias{gtkFileChooserGetType}
\alias{gtkFileChooserWidgetGetType}
\alias{gtkFileFilterGetType}
\alias{gtkFileSelectionGetType}
\alias{gtkFixedGetType}
\alias{gtkFontButtonGetType}
\alias{gtkFontSelectionDialogGetType}
\alias{gtkFontSelectionGetType}
\alias{gtkFrameGetType}
\alias{gtkGammaCurveGetType}
\alias{gtkHBoxGetType}
\alias{gtkHButtonBoxGetType}
\alias{gtkHPanedGetType}
\alias{gtkHRulerGetType}
\alias{gtkHScaleGetType}
\alias{gtkHScrollbarGetType}
\alias{gtkHSeparatorGetType}
\alias{gtkHandleBoxGetType}
\alias{gtkIMContextGetType}
\alias{gtkIMContextSimpleGetType}
\alias{gtkIMMulticontextGetType}
\alias{gtkIconFactoryGetType}
\alias{gtkIconInfoGetType}
\alias{gtkIconSetGetType}
\alias{gtkIconSourceGetType}
\alias{gtkIconThemeGetType}
\alias{gtkIconViewGetType}
\alias{gtkImageGetType}
\alias{gtkImageMenuItemGetType}
\alias{gtkInputDialogGetType}
\alias{gtkInvisibleGetType}
\alias{gtkItemFactoryGetType}
\alias{gtkItemGetType}
\alias{gtkLabelGetType}
\alias{gtkLayoutGetType}
\alias{gtkListGetType}
\alias{gtkListItemGetType}
\alias{gtkListStoreGetType}
\alias{gtkMenuBarGetType}
\alias{gtkMenuGetType}
\alias{gtkMenuItemGetType}
\alias{gtkMenuShellGetType}
\alias{gtkMenuToolButtonGetType}
\alias{gtkMessageDialogGetType}
\alias{gtkMiscGetType}
\alias{gtkNotebookGetType}
\alias{gtkOldEditableGetType}
\alias{gtkOptionMenuGetType}
\alias{gtkPanedGetType}
\alias{gtkPixmapGetType}
\alias{gtkPlugGetType}
\alias{gtkPreviewGetType}
\alias{gtkProgressBarGetType}
\alias{gtkProgressGetType}
\alias{gtkRadioActionGetType}
\alias{gtkRadioButtonGetType}
\alias{gtkRadioMenuItemGetType}
\alias{gtkRadioToolButtonGetType}
\alias{gtkRangeGetType}
\alias{gtkRcStyleGetType}
\alias{gtkRequisitionGetType}
\alias{gtkRulerGetType}
\alias{gtkScaleGetType}
\alias{gtkScrollbarGetType}
\alias{gtkScrolledWindowGetType}
\alias{gtkSelectionDataGetType}
\alias{gtkSeparatorGetType}
\alias{gtkSeparatorMenuItemGetType}
\alias{gtkSeparatorToolItemGetType}
\alias{gtkSettingsGetType}
\alias{gtkSizeGroupGetType}
\alias{gtkSocketGetType}
\alias{gtkSpinButtonGetType}
\alias{gtkStatusbarGetType}
\alias{gtkStyleGetType}
\alias{gtkTableGetType}
\alias{gtkTearoffMenuItemGetType}
\alias{gtkTextAttributesGetType}
\alias{gtkTextBufferGetType}
\alias{gtkTextChildAnchorGetType}
\alias{gtkTextIterGetType}
\alias{gtkTextMarkGetType}
\alias{gtkTextTagGetType}
\alias{gtkTextTagTableGetType}
\alias{gtkTextViewGetType}
\alias{gtkTipsQueryGetType}
\alias{gtkToggleActionGetType}
\alias{gtkToggleButtonGetType}
\alias{gtkToggleToolButtonGetType}
\alias{gtkToolButtonGetType}
\alias{gtkToolItemGetType}
\alias{gtkToolbarGetType}
\alias{gtkTooltipsGetType}
\alias{gtkTreeDragDestGetType}
\alias{gtkTreeDragSourceGetType}
\alias{gtkTreeIterGetType}
\alias{gtkTreeModelFilterGetType}
\alias{gtkTreeModelGetType}
\alias{gtkTreeModelSortGetType}
\alias{gtkTreeRowReferenceGetType}
\alias{gtkTreeSelectionGetType}
\alias{gtkTreeSortableGetType}
\alias{gtkTreeStoreGetType}
\alias{gtkTreeViewColumnGetType}
\alias{gtkTreeViewGetType}
\alias{gtkUIManagerGetType}
\alias{gtkVBoxGetType}
\alias{gtkVButtonBoxGetType}
\alias{gtkVPanedGetType}
\alias{gtkVRulerGetType}
\alias{gtkVScaleGetType}
\alias{gtkVScrollbarGetType}
\alias{gtkVSeparatorGetType}
\alias{gtkViewportGetType}
\alias{gtkWidgetGetType}
\alias{gtkWindowGetType}
\alias{gtkWindowGroupGetType}
\alias{pangoAttrClassGetType}
\alias{pangoAttrListGetType}
\alias{pangoCairoFontMapGetType}
\alias{pangoColorGetType}
\alias{pangoFontFaceGetType}
\alias{pangoFontFamilyGetType}
\alias{pangoFontGetType}
\alias{pangoFontMetricsGetType}
\alias{pangoGlyphStringGetType}
\alias{pangoLayoutGetType}
\alias{pangoLayoutIterGetType}
\alias{pangoRendererGetType}
\alias{pangoTabArrayGetType}
\name{typeMacros}
\title{GType accessor macros}
\description{This function returns the R representation of the GType of the object argument.}
\author{Michael Lawrence}
\keyword{interface}
\keyword{internal}
