\name{utils}
\alias{imagefile}
\title{RGtk2 Utilities}
\description{Some simple RGtk2 utilities that you probably don't need.}
\usage{
imagefile(name)
}
\arguments{
\item{name}{The file name to which to append the path.}
}
\details{
\code{imagefile} appends the path to the RGtk2 images directory onto the given filename.
}
\author{Michael Lawrence}
\keyword{internal}
