\alias{atkActionIfaceDoAction}
\alias{atkActionIfaceGetDescription}
\alias{atkActionIfaceGetKeybinding}
\alias{atkActionIfaceGetLocalizedName}
\alias{atkActionIfaceGetNActions}
\alias{atkActionIfaceGetName}
\alias{atkActionIfaceSetDescription}
\alias{atkComponentIfaceBoundsChanged}
\alias{atkComponentIfaceContains}
\alias{atkComponentIfaceGetExtents}
\alias{atkComponentIfaceGetLayer}
\alias{atkComponentIfaceGetMdiZorder}
\alias{atkComponentIfaceGetPosition}
\alias{atkComponentIfaceGetSize}
\alias{atkComponentIfaceGrabFocus}
\alias{atkComponentIfaceRefAccessibleAtPoint}
\alias{atkComponentIfaceRemoveFocusHandler}
\alias{atkComponentIfaceSetExtents}
\alias{atkComponentIfaceSetPosition}
\alias{atkComponentIfaceSetSize}
\alias{atkDocumentIfaceGetDocument}
\alias{atkDocumentIfaceGetDocumentType}
\alias{atkEditableTextIfaceCopyText}
\alias{atkEditableTextIfaceCutText}
\alias{atkEditableTextIfaceDeleteText}
\alias{atkEditableTextIfaceInsertText}
\alias{atkEditableTextIfacePasteText}
\alias{atkEditableTextIfaceSetRunAttributes}
\alias{atkEditableTextIfaceSetTextContents}
\alias{atkHyperlinkClassGetEndIndex}
\alias{atkHyperlinkClassGetNAnchors}
\alias{atkHyperlinkClassGetObject}
\alias{atkHyperlinkClassGetStartIndex}
\alias{atkHyperlinkClassGetUri}
\alias{atkHyperlinkClassIsSelectedLink}
\alias{atkHyperlinkClassIsValid}
\alias{atkHyperlinkClassLinkActivated}
\alias{atkHyperlinkClassLinkState}
\alias{atkHyperlinkImplIfaceGetHyperlink}
\alias{atkHypertextIfaceGetLink}
\alias{atkHypertextIfaceGetLinkIndex}
\alias{atkHypertextIfaceGetNLinks}
\alias{atkHypertextIfaceLinkSelected}
\alias{atkImageIfaceGetImageDescription}
\alias{atkImageIfaceGetImagePosition}
\alias{atkImageIfaceGetImageSize}
\alias{atkImageIfaceSetImageDescription}
\alias{atkImplementorIfaceRefAccessible}
\alias{atkObjectClassActiveDescendantChanged}
\alias{atkObjectClassChildrenChanged}
\alias{atkObjectClassFocusEvent}
\alias{atkObjectClassGetDescription}
\alias{atkObjectClassGetIndexInParent}
\alias{atkObjectClassGetLayer}
\alias{atkObjectClassGetMdiZorder}
\alias{atkObjectClassGetNChildren}
\alias{atkObjectClassGetName}
\alias{atkObjectClassGetParent}
\alias{atkObjectClassGetRole}
\alias{atkObjectClassInitialize}
\alias{atkObjectClassRefChild}
\alias{atkObjectClassRefRelationSet}
\alias{atkObjectClassRefStateSet}
\alias{atkObjectClassRemovePropertyChangeHandler}
\alias{atkObjectClassSetDescription}
\alias{atkObjectClassSetName}
\alias{atkObjectClassSetParent}
\alias{atkObjectClassSetRole}
\alias{atkObjectClassStateChange}
\alias{atkObjectClassVisibleDataChanged}
\alias{atkObjectFactoryClassInvalidate}
\alias{atkSelectionIfaceAddSelection}
\alias{atkSelectionIfaceClearSelection}
\alias{atkSelectionIfaceGetSelectionCount}
\alias{atkSelectionIfaceIsChildSelected}
\alias{atkSelectionIfaceRefSelection}
\alias{atkSelectionIfaceRemoveSelection}
\alias{atkSelectionIfaceSelectAllSelection}
\alias{atkSelectionIfaceSelectionChanged}
\alias{atkStreamableContentIfaceGetMimeType}
\alias{atkStreamableContentIfaceGetNMimeTypes}
\alias{atkTableIfaceAddColumnSelection}
\alias{atkTableIfaceAddRowSelection}
\alias{atkTableIfaceColumnDeleted}
\alias{atkTableIfaceColumnInserted}
\alias{atkTableIfaceColumnReordered}
\alias{atkTableIfaceGetCaption}
\alias{atkTableIfaceGetColumnAtIndex}
\alias{atkTableIfaceGetColumnDescription}
\alias{atkTableIfaceGetColumnExtentAt}
\alias{atkTableIfaceGetColumnHeader}
\alias{atkTableIfaceGetIndexAt}
\alias{atkTableIfaceGetNColumns}
\alias{atkTableIfaceGetNRows}
\alias{atkTableIfaceGetRowAtIndex}
\alias{atkTableIfaceGetRowDescription}
\alias{atkTableIfaceGetRowExtentAt}
\alias{atkTableIfaceGetRowHeader}
\alias{atkTableIfaceGetSelectedColumns}
\alias{atkTableIfaceGetSelectedRows}
\alias{atkTableIfaceGetSummary}
\alias{atkTableIfaceIsColumnSelected}
\alias{atkTableIfaceIsRowSelected}
\alias{atkTableIfaceIsSelected}
\alias{atkTableIfaceModelChanged}
\alias{atkTableIfaceRefAt}
\alias{atkTableIfaceRemoveColumnSelection}
\alias{atkTableIfaceRemoveRowSelection}
\alias{atkTableIfaceRowDeleted}
\alias{atkTableIfaceRowInserted}
\alias{atkTableIfaceRowReordered}
\alias{atkTableIfaceSetCaption}
\alias{atkTableIfaceSetColumnDescription}
\alias{atkTableIfaceSetColumnHeader}
\alias{atkTableIfaceSetRowDescription}
\alias{atkTableIfaceSetRowHeader}
\alias{atkTableIfaceSetSummary}
\alias{atkTextIfaceAddSelection}
\alias{atkTextIfaceGetBoundedRanges}
\alias{atkTextIfaceGetCaretOffset}
\alias{atkTextIfaceGetCharacterAtOffset}
\alias{atkTextIfaceGetCharacterCount}
\alias{atkTextIfaceGetCharacterExtents}
\alias{atkTextIfaceGetDefaultAttributes}
\alias{atkTextIfaceGetNSelections}
\alias{atkTextIfaceGetOffsetAtPoint}
\alias{atkTextIfaceGetRangeExtents}
\alias{atkTextIfaceGetRunAttributes}
\alias{atkTextIfaceGetSelection}
\alias{atkTextIfaceGetText}
\alias{atkTextIfaceGetTextAfterOffset}
\alias{atkTextIfaceGetTextAtOffset}
\alias{atkTextIfaceGetTextBeforeOffset}
\alias{atkTextIfaceRemoveSelection}
\alias{atkTextIfaceSetCaretOffset}
\alias{atkTextIfaceSetSelection}
\alias{atkTextIfaceTextAttributesChanged}
\alias{atkTextIfaceTextCaretMoved}
\alias{atkTextIfaceTextChanged}
\alias{atkTextIfaceTextSelectionChanged}
\alias{atkValueIfaceGetCurrentValue}
\alias{atkValueIfaceGetMaximumValue}
\alias{atkValueIfaceGetMinimumIncrement}
\alias{atkValueIfaceGetMinimumValue}
\alias{atkValueIfaceSetCurrentValue}
\alias{gdkDisplayClassClosed}
\alias{gdkDisplayClassGetDefaultScreen}
\alias{gdkDisplayClassGetDisplayName}
\alias{gdkDisplayClassGetNScreens}
\alias{gdkDisplayClassGetScreen}
\alias{gdkDisplayManagerClassDisplayOpened}
\alias{gdkDrawableClassCreateGc}
\alias{gdkDrawableClassDrawArc}
\alias{gdkDrawableClassDrawDrawable}
\alias{gdkDrawableClassDrawGlyphs}
\alias{gdkDrawableClassDrawGlyphsTransformed}
\alias{gdkDrawableClassDrawImage}
\alias{gdkDrawableClassDrawLines}
\alias{gdkDrawableClassDrawPixbuf}
\alias{gdkDrawableClassDrawPoints}
\alias{gdkDrawableClassDrawPolygon}
\alias{gdkDrawableClassDrawRectangle}
\alias{gdkDrawableClassDrawSegments}
\alias{gdkDrawableClassDrawText}
\alias{gdkDrawableClassDrawTextWc}
\alias{gdkDrawableClassDrawTrapezoids}
\alias{gdkDrawableClassGetClipRegion}
\alias{gdkDrawableClassGetColormap}
\alias{gdkDrawableClassGetCompositeDrawable}
\alias{gdkDrawableClassGetDepth}
\alias{gdkDrawableClassGetImage}
\alias{gdkDrawableClassGetScreen}
\alias{gdkDrawableClassGetSize}
\alias{gdkDrawableClassGetVisibleRegion}
\alias{gdkDrawableClassGetVisual}
\alias{gdkDrawableClassRefCairoSurface}
\alias{gdkDrawableClassSetColormap}
\alias{gdkGCClassGetValues}
\alias{gdkGCClassSetDashes}
\alias{gdkGCClassSetValues}
\alias{gdkKeymapClassDirectionChanged}
\alias{gdkKeymapClassKeysChanged}
\alias{gdkPixbufAnimationClassGetIter}
\alias{gdkPixbufAnimationClassGetSize}
\alias{gdkPixbufAnimationClassGetStaticImage}
\alias{gdkPixbufAnimationClassIsStaticImage}
\alias{gdkPixbufAnimationIterClassAdvance}
\alias{gdkPixbufAnimationIterClassGetDelayTime}
\alias{gdkPixbufAnimationIterClassGetPixbuf}
\alias{gdkPixbufAnimationIterClassOnCurrentlyLoadingFrame}
\alias{gdkPixbufLoaderClassAreaPrepared}
\alias{gdkPixbufLoaderClassAreaUpdated}
\alias{gdkPixbufLoaderClassClosed}
\alias{gdkPixbufLoaderClassSizePrepared}
\alias{gdkScreenClassCompositedChanged}
\alias{gdkScreenClassSizeChanged}
\alias{gtkAccelGroupClassAccelChanged}
\alias{gtkAccessibleClassConnectWidgetDestroyed}
\alias{gtkActionClassActivate}
\alias{gtkActionClassConnectProxy}
\alias{gtkActionClassCreateMenuItem}
\alias{gtkActionClassCreateToolItem}
\alias{gtkActionClassDisconnectProxy}
\alias{gtkActionGroupClassGetAction}
\alias{gtkAdjustmentClassChanged}
\alias{gtkAdjustmentClassValueChanged}
\alias{gtkAssistantClassApply}
\alias{gtkAssistantClassCancel}
\alias{gtkAssistantClassClose}
\alias{gtkAssistantClassPrepare}
\alias{gtkButtonClassActivate}
\alias{gtkButtonClassClicked}
\alias{gtkButtonClassEnter}
\alias{gtkButtonClassLeave}
\alias{gtkButtonClassPressed}
\alias{gtkButtonClassReleased}
\alias{gtkCListClassAbortColumnResize}
\alias{gtkCListClassCellSizeRequest}
\alias{gtkCListClassClear}
\alias{gtkCListClassClickColumn}
\alias{gtkCListClassDrawDragHighlight}
\alias{gtkCListClassDrawRow}
\alias{gtkCListClassEndSelection}
\alias{gtkCListClassExtendSelection}
\alias{gtkCListClassFakeUnselectAll}
\alias{gtkCListClassInsertRow}
\alias{gtkCListClassRefresh}
\alias{gtkCListClassRemoveRow}
\alias{gtkCListClassResizeColumn}
\alias{gtkCListClassResyncSelection}
\alias{gtkCListClassRowMove}
\alias{gtkCListClassScrollHorizontal}
\alias{gtkCListClassScrollVertical}
\alias{gtkCListClassSelectAll}
\alias{gtkCListClassSelectRow}
\alias{gtkCListClassSelectionFind}
\alias{gtkCListClassSetCellContents}
\alias{gtkCListClassSetScrollAdjustments}
\alias{gtkCListClassSortList}
\alias{gtkCListClassStartSelection}
\alias{gtkCListClassToggleAddMode}
\alias{gtkCListClassToggleFocusRow}
\alias{gtkCListClassUndoSelection}
\alias{gtkCListClassUnselectAll}
\alias{gtkCListClassUnselectRow}
\alias{gtkCTreeClassChangeFocusRowExpansion}
\alias{gtkCTreeClassTreeCollapse}
\alias{gtkCTreeClassTreeExpand}
\alias{gtkCTreeClassTreeMove}
\alias{gtkCTreeClassTreeSelectRow}
\alias{gtkCTreeClassTreeUnselectRow}
\alias{gtkCalendarClassDaySelected}
\alias{gtkCalendarClassDaySelectedDoubleClick}
\alias{gtkCalendarClassMonthChanged}
\alias{gtkCalendarClassNextMonth}
\alias{gtkCalendarClassNextYear}
\alias{gtkCalendarClassPrevMonth}
\alias{gtkCalendarClassPrevYear}
\alias{gtkCellEditableIfaceEditingDone}
\alias{gtkCellEditableIfaceRemoveWidget}
\alias{gtkCellEditableIfaceStartEditing}
\alias{gtkCellLayoutIfaceAddAttribute}
\alias{gtkCellLayoutIfaceClear}
\alias{gtkCellLayoutIfaceClearAttributes}
\alias{gtkCellLayoutIfacePackEnd}
\alias{gtkCellLayoutIfacePackStart}
\alias{gtkCellLayoutIfaceReorder}
\alias{gtkCellLayoutIfaceSetCellDataFunc}
\alias{gtkCellRendererAccelClassAccelCleared}
\alias{gtkCellRendererAccelClassAccelEdited}
\alias{gtkCellRendererClassActivate}
\alias{gtkCellRendererClassEditingCanceled}
\alias{gtkCellRendererClassEditingStarted}
\alias{gtkCellRendererClassGetSize}
\alias{gtkCellRendererClassRender}
\alias{gtkCellRendererClassStartEditing}
\alias{gtkCellRendererTextClassEdited}
\alias{gtkCellRendererToggleClassToggled}
\alias{gtkCheckButtonClassDrawIndicator}
\alias{gtkCheckMenuItemClassDrawIndicator}
\alias{gtkCheckMenuItemClassToggled}
\alias{gtkColorButtonClassColorSet}
\alias{gtkColorSelectionClassColorChanged}
\alias{gtkComboBoxClassChanged}
\alias{gtkComboBoxClassGetActiveText}
\alias{gtkContainerClassAdd}
\alias{gtkContainerClassCheckResize}
\alias{gtkContainerClassChildType}
\alias{gtkContainerClassCompositeName}
\alias{gtkContainerClassForall}
\alias{gtkContainerClassGetChildProperty}
\alias{gtkContainerClassRemove}
\alias{gtkContainerClassSetChildProperty}
\alias{gtkContainerClassSetFocusChild}
\alias{gtkCurveClassCurveTypeChanged}
\alias{gtkDialogClassClose}
\alias{gtkDialogClassResponse}
\alias{gtkEditableIfaceChanged}
\alias{gtkEditableIfaceDeleteText}
\alias{gtkEditableIfaceDoDeleteText}
\alias{gtkEditableIfaceDoInsertText}
\alias{gtkEditableIfaceGetChars}
\alias{gtkEditableIfaceGetPosition}
\alias{gtkEditableIfaceGetSelectionBounds}
\alias{gtkEditableIfaceInsertText}
\alias{gtkEditableIfaceSetPosition}
\alias{gtkEditableIfaceSetSelectionBounds}
\alias{gtkEntryClassActivate}
\alias{gtkEntryClassBackspace}
\alias{gtkEntryClassCopyClipboard}
\alias{gtkEntryClassCutClipboard}
\alias{gtkEntryClassDeleteFromCursor}
\alias{gtkEntryClassInsertAtCursor}
\alias{gtkEntryClassMoveCursor}
\alias{gtkEntryClassPasteClipboard}
\alias{gtkEntryClassPopulatePopup}
\alias{gtkEntryClassToggleOverwrite}
\alias{gtkEntryCompletionClassActionActivated}
\alias{gtkEntryCompletionClassInsertPrefix}
\alias{gtkEntryCompletionClassMatchSelected}
\alias{gtkExpanderClassActivate}
\alias{gtkFontButtonClassFontSet}
\alias{gtkFrameClassComputeChildAllocation}
\alias{gtkHandleBoxClassChildAttached}
\alias{gtkHandleBoxClassChildDetached}
\alias{gtkIMContextClassCommit}
\alias{gtkIMContextClassDeleteSurrounding}
\alias{gtkIMContextClassFilterKeypress}
\alias{gtkIMContextClassFocusIn}
\alias{gtkIMContextClassFocusOut}
\alias{gtkIMContextClassGetPreeditString}
\alias{gtkIMContextClassGetSurrounding}
\alias{gtkIMContextClassPreeditChanged}
\alias{gtkIMContextClassPreeditEnd}
\alias{gtkIMContextClassPreeditStart}
\alias{gtkIMContextClassReset}
\alias{gtkIMContextClassRetrieveSurrounding}
\alias{gtkIMContextClassSetClientWindow}
\alias{gtkIMContextClassSetCursorLocation}
\alias{gtkIMContextClassSetSurrounding}
\alias{gtkIMContextClassSetUsePreedit}
\alias{gtkIconThemeClassChanged}
\alias{gtkIconViewClassActivateCursorItem}
\alias{gtkIconViewClassItemActivated}
\alias{gtkIconViewClassMoveCursor}
\alias{gtkIconViewClassSelectAll}
\alias{gtkIconViewClassSelectCursorItem}
\alias{gtkIconViewClassSelectionChanged}
\alias{gtkIconViewClassSetScrollAdjustments}
\alias{gtkIconViewClassToggleCursorItem}
\alias{gtkIconViewClassUnselectAll}
\alias{gtkInputDialogClassDisableDevice}
\alias{gtkInputDialogClassEnableDevice}
\alias{gtkItemClassDeselect}
\alias{gtkItemClassSelect}
\alias{gtkItemClassToggle}
\alias{gtkLabelClassCopyClipboard}
\alias{gtkLabelClassMoveCursor}
\alias{gtkLabelClassPopulatePopup}
\alias{gtkLayoutClassSetScrollAdjustments}
\alias{gtkListClassSelectChild}
\alias{gtkListClassSelectionChanged}
\alias{gtkListClassUnselectChild}
\alias{gtkListItemClassEndSelection}
\alias{gtkListItemClassExtendSelection}
\alias{gtkListItemClassScrollHorizontal}
\alias{gtkListItemClassScrollVertical}
\alias{gtkListItemClassSelectAll}
\alias{gtkListItemClassStartSelection}
\alias{gtkListItemClassToggleAddMode}
\alias{gtkListItemClassToggleFocusRow}
\alias{gtkListItemClassUndoSelection}
\alias{gtkListItemClassUnselectAll}
\alias{gtkMenuItemClassActivate}
\alias{gtkMenuItemClassActivateItem}
\alias{gtkMenuItemClassToggleSizeAllocate}
\alias{gtkMenuItemClassToggleSizeRequest}
\alias{gtkMenuShellClassActivateCurrent}
\alias{gtkMenuShellClassCancel}
\alias{gtkMenuShellClassDeactivate}
\alias{gtkMenuShellClassGetPopupDelay}
\alias{gtkMenuShellClassInsert}
\alias{gtkMenuShellClassMoveCurrent}
\alias{gtkMenuShellClassSelectItem}
\alias{gtkMenuShellClassSelectionDone}
\alias{gtkMenuToolButtonClassShowMenu}
\alias{gtkNotebookClassChangeCurrentPage}
\alias{gtkNotebookClassFocusTab}
\alias{gtkNotebookClassInsertPage}
\alias{gtkNotebookClassMoveFocusOut}
\alias{gtkNotebookClassReorderTab}
\alias{gtkNotebookClassSelectPage}
\alias{gtkNotebookClassSwitchPage}
\alias{gtkOldEditableClassActivate}
\alias{gtkOldEditableClassCopyClipboard}
\alias{gtkOldEditableClassCutClipboard}
\alias{gtkOldEditableClassGetChars}
\alias{gtkOldEditableClassKillChar}
\alias{gtkOldEditableClassKillLine}
\alias{gtkOldEditableClassKillWord}
\alias{gtkOldEditableClassMoveCursor}
\alias{gtkOldEditableClassMovePage}
\alias{gtkOldEditableClassMoveToColumn}
\alias{gtkOldEditableClassMoveToRow}
\alias{gtkOldEditableClassMoveWord}
\alias{gtkOldEditableClassPasteClipboard}
\alias{gtkOldEditableClassSetEditable}
\alias{gtkOldEditableClassSetPosition}
\alias{gtkOldEditableClassSetSelection}
\alias{gtkOldEditableClassUpdateText}
\alias{gtkOptionMenuClassChanged}
\alias{gtkPanedClassAcceptPosition}
\alias{gtkPanedClassCancelPosition}
\alias{gtkPanedClassCycleChildFocus}
\alias{gtkPanedClassCycleHandleFocus}
\alias{gtkPanedClassMoveHandle}
\alias{gtkPanedClassToggleHandleFocus}
\alias{gtkPlugClassEmbedded}
\alias{gtkPrintOperationClassBeginPrint}
\alias{gtkPrintOperationClassCreateCustomWidget}
\alias{gtkPrintOperationClassCustomWidgetApply}
\alias{gtkPrintOperationClassDone}
\alias{gtkPrintOperationClassDrawPage}
\alias{gtkPrintOperationClassEndPrint}
\alias{gtkPrintOperationClassPaginate}
\alias{gtkPrintOperationClassPreview}
\alias{gtkPrintOperationClassRequestPageSetup}
\alias{gtkPrintOperationClassStatusChanged}
\alias{gtkPrintOperationPreviewClassEndPreview}
\alias{gtkPrintOperationPreviewClassGotPageSize}
\alias{gtkPrintOperationPreviewClassIsSelected}
\alias{gtkPrintOperationPreviewClassReady}
\alias{gtkPrintOperationPreviewClassRenderPage}
\alias{gtkProgressClassActModeEnter}
\alias{gtkProgressClassPaint}
\alias{gtkProgressClassUpdate}
\alias{gtkRadioActionClassChanged}
\alias{gtkRadioButtonClassGroupChanged}
\alias{gtkRadioMenuItemClassGroupChanged}
\alias{gtkRangeClassAdjustBounds}
\alias{gtkRangeClassChangeValue}
\alias{gtkRangeClassGetRangeBorder}
\alias{gtkRangeClassMoveSlider}
\alias{gtkRangeClassValueChanged}
\alias{gtkRcStyleClassCreateRcStyle}
\alias{gtkRcStyleClassCreateStyle}
\alias{gtkRcStyleClassMerge}
\alias{gtkRcStyleClassParse}
\alias{gtkRecentChooserClassAddFilter}
\alias{gtkRecentChooserClassGetCurrentUri}
\alias{gtkRecentChooserClassGetItems}
\alias{gtkRecentChooserClassGetRecentManager}
\alias{gtkRecentChooserClassItemActivated}
\alias{gtkRecentChooserClassListFilters}
\alias{gtkRecentChooserClassRemoveFilter}
\alias{gtkRecentChooserClassSelectAll}
\alias{gtkRecentChooserClassSelectUri}
\alias{gtkRecentChooserClassSelectionChanged}
\alias{gtkRecentChooserClassSetCurrentUri}
\alias{gtkRecentChooserClassSetSortFunc}
\alias{gtkRecentChooserClassUnselectAll}
\alias{gtkRecentChooserClassUnselectUri}
\alias{gtkRecentManagerClassChanged}
\alias{gtkRulerClassDrawPos}
\alias{gtkRulerClassDrawTicks}
\alias{gtkScaleClassDrawValue}
\alias{gtkScaleClassFormatValue}
\alias{gtkScaleClassGetLayoutOffsets}
\alias{gtkScrolledWindowClassMoveFocusOut}
\alias{gtkScrolledWindowClassScrollChild}
\alias{gtkSocketClassPlugAdded}
\alias{gtkSocketClassPlugRemoved}
\alias{gtkSpinButtonClassChangeValue}
\alias{gtkSpinButtonClassInput}
\alias{gtkSpinButtonClassOutput}
\alias{gtkSpinButtonClassValueChanged}
\alias{gtkSpinButtonClassWrapped}
\alias{gtkStatusIconClassActivate}
\alias{gtkStatusIconClassPopupMenu}
\alias{gtkStatusIconClassSizeChanged}
\alias{gtkStatusbarClassTextPopped}
\alias{gtkStatusbarClassTextPushed}
\alias{gtkStyleClassClone}
\alias{gtkStyleClassCopy}
\alias{gtkStyleClassDrawArrow}
\alias{gtkStyleClassDrawBox}
\alias{gtkStyleClassDrawBoxGap}
\alias{gtkStyleClassDrawCheck}
\alias{gtkStyleClassDrawDiamond}
\alias{gtkStyleClassDrawExpander}
\alias{gtkStyleClassDrawExtension}
\alias{gtkStyleClassDrawFlatBox}
\alias{gtkStyleClassDrawFocus}
\alias{gtkStyleClassDrawHandle}
\alias{gtkStyleClassDrawHline}
\alias{gtkStyleClassDrawLayout}
\alias{gtkStyleClassDrawOption}
\alias{gtkStyleClassDrawPolygon}
\alias{gtkStyleClassDrawResizeGrip}
\alias{gtkStyleClassDrawShadow}
\alias{gtkStyleClassDrawShadowGap}
\alias{gtkStyleClassDrawSlider}
\alias{gtkStyleClassDrawString}
\alias{gtkStyleClassDrawTab}
\alias{gtkStyleClassDrawVline}
\alias{gtkStyleClassInitFromRc}
\alias{gtkStyleClassRealize}
\alias{gtkStyleClassRenderIcon}
\alias{gtkStyleClassSetBackground}
\alias{gtkStyleClassUnrealize}
\alias{gtkTextBufferClassApplyTag}
\alias{gtkTextBufferClassBeginUserAction}
\alias{gtkTextBufferClassChanged}
\alias{gtkTextBufferClassDeleteRange}
\alias{gtkTextBufferClassEndUserAction}
\alias{gtkTextBufferClassInsertChildAnchor}
\alias{gtkTextBufferClassInsertPixbuf}
\alias{gtkTextBufferClassInsertText}
\alias{gtkTextBufferClassMarkDeleted}
\alias{gtkTextBufferClassMarkSet}
\alias{gtkTextBufferClassModifiedChanged}
\alias{gtkTextBufferClassRemoveTag}
\alias{gtkTextClassSetScrollAdjustments}
\alias{gtkTextTagClassEvent}
\alias{gtkTextTagTableClassTagAdded}
\alias{gtkTextTagTableClassTagChanged}
\alias{gtkTextTagTableClassTagRemoved}
\alias{gtkTextViewClassBackspace}
\alias{gtkTextViewClassCopyClipboard}
\alias{gtkTextViewClassCutClipboard}
\alias{gtkTextViewClassDeleteFromCursor}
\alias{gtkTextViewClassInsertAtCursor}
\alias{gtkTextViewClassMoveCursor}
\alias{gtkTextViewClassMoveFocus}
\alias{gtkTextViewClassPageHorizontally}
\alias{gtkTextViewClassPasteClipboard}
\alias{gtkTextViewClassPopulatePopup}
\alias{gtkTextViewClassSetAnchor}
\alias{gtkTextViewClassSetScrollAdjustments}
\alias{gtkTextViewClassToggleOverwrite}
\alias{gtkTipsQueryClassStartQuery}
\alias{gtkTipsQueryClassStopQuery}
\alias{gtkTipsQueryClassWidgetEntered}
\alias{gtkTipsQueryClassWidgetSelected}
\alias{gtkToggleActionClassToggled}
\alias{gtkToggleButtonClassToggled}
\alias{gtkToggleToolButtonClassToggled}
\alias{gtkToolButtonClassClicked}
\alias{gtkToolItemClassCreateMenuProxy}
\alias{gtkToolItemClassSetTooltip}
\alias{gtkToolItemClassToolbarReconfigured}
\alias{gtkToolbarClassOrientationChanged}
\alias{gtkToolbarClassPopupContextMenu}
\alias{gtkToolbarClassStyleChanged}
\alias{gtkTreeClassSelectChild}
\alias{gtkTreeClassUnselectChild}
\alias{gtkTreeDragDestIfaceDragDataReceived}
\alias{gtkTreeDragDestIfaceRowDropPossible}
\alias{gtkTreeDragSourceIfaceDragDataDelete}
\alias{gtkTreeDragSourceIfaceDragDataGet}
\alias{gtkTreeDragSourceIfaceRowDraggable}
\alias{gtkTreeItemClassCollapse}
\alias{gtkTreeItemClassExpand}
\alias{gtkTreeModelIfaceGetColumnType}
\alias{gtkTreeModelIfaceGetFlags}
\alias{gtkTreeModelIfaceGetIter}
\alias{gtkTreeModelIfaceGetNColumns}
\alias{gtkTreeModelIfaceGetPath}
\alias{gtkTreeModelIfaceGetValue}
\alias{gtkTreeModelIfaceIterChildren}
\alias{gtkTreeModelIfaceIterHasChild}
\alias{gtkTreeModelIfaceIterNChildren}
\alias{gtkTreeModelIfaceIterNext}
\alias{gtkTreeModelIfaceIterNthChild}
\alias{gtkTreeModelIfaceIterParent}
\alias{gtkTreeModelIfaceRefNode}
\alias{gtkTreeModelIfaceRowChanged}
\alias{gtkTreeModelIfaceRowDeleted}
\alias{gtkTreeModelIfaceRowHasChildToggled}
\alias{gtkTreeModelIfaceRowInserted}
\alias{gtkTreeModelIfaceRowsReordered}
\alias{gtkTreeModelIfaceUnrefNode}
\alias{gtkTreeSelectionClassChanged}
\alias{gtkTreeSortableIfaceGetSortColumnId}
\alias{gtkTreeSortableIfaceHasDefaultSortFunc}
\alias{gtkTreeSortableIfaceSetDefaultSortFunc}
\alias{gtkTreeSortableIfaceSetSortColumnId}
\alias{gtkTreeSortableIfaceSetSortFunc}
\alias{gtkTreeSortableIfaceSortColumnChanged}
\alias{gtkTreeViewClassColumnsChanged}
\alias{gtkTreeViewClassCursorChanged}
\alias{gtkTreeViewClassExpandCollapseCursorRow}
\alias{gtkTreeViewClassMoveCursor}
\alias{gtkTreeViewClassRowActivated}
\alias{gtkTreeViewClassRowCollapsed}
\alias{gtkTreeViewClassRowExpanded}
\alias{gtkTreeViewClassSelectAll}
\alias{gtkTreeViewClassSelectCursorParent}
\alias{gtkTreeViewClassSelectCursorRow}
\alias{gtkTreeViewClassSetScrollAdjustments}
\alias{gtkTreeViewClassStartInteractiveSearch}
\alias{gtkTreeViewClassTestCollapseRow}
\alias{gtkTreeViewClassTestExpandRow}
\alias{gtkTreeViewClassToggleCursorRow}
\alias{gtkTreeViewClassUnselectAll}
\alias{gtkTreeViewColumnClassClicked}
\alias{gtkUIManagerClassActionsChanged}
\alias{gtkUIManagerClassAddWidget}
\alias{gtkUIManagerClassConnectProxy}
\alias{gtkUIManagerClassDisconnectProxy}
\alias{gtkUIManagerClassGetAction}
\alias{gtkUIManagerClassGetWidget}
\alias{gtkUIManagerClassPostActivate}
\alias{gtkUIManagerClassPreActivate}
\alias{gtkViewportClassSetScrollAdjustments}
\alias{gtkWidgetClassButtonPressEvent}
\alias{gtkWidgetClassButtonReleaseEvent}
\alias{gtkWidgetClassCanActivateAccel}
\alias{gtkWidgetClassChildNotify}
\alias{gtkWidgetClassClientEvent}
\alias{gtkWidgetClassCompositedChanged}
\alias{gtkWidgetClassConfigureEvent}
\alias{gtkWidgetClassDeleteEvent}
\alias{gtkWidgetClassDestroyEvent}
\alias{gtkWidgetClassDirectionChanged}
\alias{gtkWidgetClassDispatchChildPropertiesChanged}
\alias{gtkWidgetClassDragBegin}
\alias{gtkWidgetClassDragDataDelete}
\alias{gtkWidgetClassDragDataGet}
\alias{gtkWidgetClassDragDataReceived}
\alias{gtkWidgetClassDragDrop}
\alias{gtkWidgetClassDragEnd}
\alias{gtkWidgetClassDragLeave}
\alias{gtkWidgetClassDragMotion}
\alias{gtkWidgetClassEnterNotifyEvent}
\alias{gtkWidgetClassEvent}
\alias{gtkWidgetClassExposeEvent}
\alias{gtkWidgetClassFocus}
\alias{gtkWidgetClassFocusInEvent}
\alias{gtkWidgetClassFocusOutEvent}
\alias{gtkWidgetClassGetAccessible}
\alias{gtkWidgetClassGrabBrokenEvent}
\alias{gtkWidgetClassGrabFocus}
\alias{gtkWidgetClassGrabNotify}
\alias{gtkWidgetClassHide}
\alias{gtkWidgetClassHideAll}
\alias{gtkWidgetClassHierarchyChanged}
\alias{gtkWidgetClassKeyPressEvent}
\alias{gtkWidgetClassKeyReleaseEvent}
\alias{gtkWidgetClassLeaveNotifyEvent}
\alias{gtkWidgetClassMap}
\alias{gtkWidgetClassMapEvent}
\alias{gtkWidgetClassMnemonicActivate}
\alias{gtkWidgetClassMotionNotifyEvent}
\alias{gtkWidgetClassNoExposeEvent}
\alias{gtkWidgetClassParentSet}
\alias{gtkWidgetClassPopupMenu}
\alias{gtkWidgetClassPropertyNotifyEvent}
\alias{gtkWidgetClassProximityInEvent}
\alias{gtkWidgetClassProximityOutEvent}
\alias{gtkWidgetClassRealize}
\alias{gtkWidgetClassScreenChanged}
\alias{gtkWidgetClassScrollEvent}
\alias{gtkWidgetClassSelectionClearEvent}
\alias{gtkWidgetClassSelectionGet}
\alias{gtkWidgetClassSelectionNotifyEvent}
\alias{gtkWidgetClassSelectionReceived}
\alias{gtkWidgetClassSelectionRequestEvent}
\alias{gtkWidgetClassShow}
\alias{gtkWidgetClassShowAll}
\alias{gtkWidgetClassShowHelp}
\alias{gtkWidgetClassSizeAllocate}
\alias{gtkWidgetClassSizeRequest}
\alias{gtkWidgetClassStateChanged}
\alias{gtkWidgetClassStyleSet}
\alias{gtkWidgetClassUnmap}
\alias{gtkWidgetClassUnmapEvent}
\alias{gtkWidgetClassUnrealize}
\alias{gtkWidgetClassVisibilityNotifyEvent}
\alias{gtkWidgetClassWindowStateEvent}
\alias{gtkWindowClassActivateDefault}
\alias{gtkWindowClassActivateFocus}
\alias{gtkWindowClassFrameEvent}
\alias{gtkWindowClassKeysChanged}
\alias{gtkWindowClassMoveFocus}
\alias{gtkWindowClassSetFocus}
\alias{pangoAttrClassCopy}
\alias{pangoAttrClassDestroy}
\alias{pangoAttrClassEqual}
\alias{pangoFontClassDescribe}
\alias{pangoFontClassGetCoverage}
\alias{pangoFontClassGetFontMap}
\alias{pangoFontClassGetGlyphExtents}
\alias{pangoFontClassGetMetrics}
\alias{pangoFontFaceClassDescribe}
\alias{pangoFontFaceClassGetFaceName}
\alias{pangoFontFaceClassListSizes}
\alias{pangoFontFamilyClassGetName}
\alias{pangoFontFamilyClassIsMonospace}
\alias{pangoFontFamilyClassListFaces}
\alias{pangoFontMapClassListFamilies}
\alias{pangoFontMapClassLoadFont}
\alias{pangoFontMapClassLoadFontset}
\alias{pangoFontsetClassForeach}
\alias{pangoFontsetClassGetFont}
\alias{pangoFontsetClassGetLanguage}
\alias{pangoFontsetClassGetMetrics}
\alias{pangoRendererClassBegin}
\alias{pangoRendererClassDrawErrorUnderline}
\alias{pangoRendererClassDrawGlyph}
\alias{pangoRendererClassDrawGlyphs}
\alias{pangoRendererClassDrawRectangle}
\alias{pangoRendererClassDrawShape}
\alias{pangoRendererClassDrawTrapezoid}
\alias{pangoRendererClassEnd}
\alias{pangoRendererClassPartChanged}
\alias{pangoRendererClassPrepareRun}
\alias{gdkGCclassGetValues}
\alias{gdkGCclassSetDashes}
\alias{gdkGCclassSetValues}
\alias{gtkBuildableIfaceAddChild}
\alias{gtkBuildableIfaceConstructChild}
\alias{gtkBuildableIfaceCustomFinished}
\alias{gtkBuildableIfaceCustomTagEnd}
\alias{gtkBuildableIfaceCustomTagStart}
\alias{gtkBuildableIfaceGetInternalChild}
\alias{gtkBuildableIfaceGetName}
\alias{gtkBuildableIfaceParserFinished}
\alias{gtkBuildableIfaceSetBuildableProperty}
\alias{gtkBuildableIfaceSetName}
\alias{gtkBuilderClassGetTypeFromName}
\alias{gAppInfoIfaceAddSupportsType}
\alias{gAppInfoIfaceCanRemoveSupportsType}
\alias{gAppInfoIfaceDup}
\alias{gAppInfoIfaceEqual}
\alias{gAppInfoIfaceGetCommandline}
\alias{gAppInfoIfaceGetDescription}
\alias{gAppInfoIfaceGetExecutable}
\alias{gAppInfoIfaceGetIcon}
\alias{gAppInfoIfaceGetId}
\alias{gAppInfoIfaceGetName}
\alias{gAppInfoIfaceLaunch}
\alias{gAppInfoIfaceLaunchUris}
\alias{gAppInfoIfaceRemoveSupportsType}
\alias{gAppInfoIfaceSetAsDefaultForExtension}
\alias{gAppInfoIfaceSetAsDefaultForType}
\alias{gAppInfoIfaceShouldShow}
\alias{gAppInfoIfaceSupportsFiles}
\alias{gAppInfoIfaceSupportsUris}
\alias{gAppLaunchContextClassGetDisplay}
\alias{gAppLaunchContextClassGetStartupNotifyId}
\alias{gAppLaunchContextClassLaunchFailed}
\alias{gAsyncInitableIfaceInitAsync}
\alias{gAsyncInitableIfaceInitFinish}
\alias{gAsyncResultIfaceGetSourceObject}
\alias{gAsyncResultIfaceGetUserData}
\alias{gBufferedInputStreamClassFill}
\alias{gBufferedInputStreamClassFillAsync}
\alias{gBufferedInputStreamClassFillFinish}
\alias{gDriveIfaceCanEject}
\alias{gDriveIfaceCanPollForMedia}
\alias{gDriveIfaceCanStart}
\alias{gDriveIfaceCanStartDegraded}
\alias{gDriveIfaceCanStop}
\alias{gDriveIfaceEject}
\alias{gDriveIfaceEjectFinish}
\alias{gDriveIfaceEjectWithOperation}
\alias{gDriveIfaceEjectWithOperationFinish}
\alias{gDriveIfaceEnumerateIdentifiers}
\alias{gDriveIfaceGetIcon}
\alias{gDriveIfaceGetIdentifier}
\alias{gDriveIfaceGetName}
\alias{gDriveIfaceGetStartStopType}
\alias{gDriveIfaceGetVolumes}
\alias{gDriveIfaceHasMedia}
\alias{gDriveIfaceHasVolumes}
\alias{gDriveIfaceIsMediaCheckAutomatic}
\alias{gDriveIfaceIsMediaRemovable}
\alias{gDriveIfacePollForMedia}
\alias{gDriveIfacePollForMediaFinish}
\alias{gDriveIfaceStart}
\alias{gDriveIfaceStartFinish}
\alias{gDriveIfaceStop}
\alias{gDriveIfaceStopFinish}
\alias{gFileEnumeratorClassCloseAsync}
\alias{gFileEnumeratorClassCloseFinish}
\alias{gFileEnumeratorClassCloseFn}
\alias{gFileEnumeratorClassNextFile}
\alias{gFileEnumeratorClassNextFilesAsync}
\alias{gFileEnumeratorClassNextFilesFinish}
\alias{gFileIOStreamClassGetEtag}
\alias{gFileIOStreamClassQueryInfo}
\alias{gFileIOStreamClassQueryInfoAsync}
\alias{gFileIOStreamClassQueryInfoFinish}
\alias{gFileIfaceAppendTo}
\alias{gFileIfaceAppendToAsync}
\alias{gFileIfaceAppendToFinish}
\alias{gFileIfaceCopy}
\alias{gFileIfaceCopyAsync}
\alias{gFileIfaceCopyFinish}
\alias{gFileIfaceCreate}
\alias{gFileIfaceCreateAsync}
\alias{gFileIfaceCreateFinish}
\alias{gFileIfaceCreateReadwrite}
\alias{gFileIfaceCreateReadwriteAsync}
\alias{gFileIfaceCreateReadwriteFinish}
\alias{gFileIfaceDeleteFile}
\alias{gFileIfaceDup}
\alias{gFileIfaceEjectMountable}
\alias{gFileIfaceEjectMountableFinish}
\alias{gFileIfaceEjectMountableWithOperation}
\alias{gFileIfaceEjectMountableWithOperationFinish}
\alias{gFileIfaceEnumerateChildren}
\alias{gFileIfaceEnumerateChildrenAsync}
\alias{gFileIfaceEnumerateChildrenFinish}
\alias{gFileIfaceEqual}
\alias{gFileIfaceFindEnclosingMount}
\alias{gFileIfaceFindEnclosingMountAsync}
\alias{gFileIfaceFindEnclosingMountFinish}
\alias{gFileIfaceGetBasename}
\alias{gFileIfaceGetChildForDisplayName}
\alias{gFileIfaceGetParent}
\alias{gFileIfaceGetParseName}
\alias{gFileIfaceGetPath}
\alias{gFileIfaceGetRelativePath}
\alias{gFileIfaceGetUri}
\alias{gFileIfaceGetUriScheme}
\alias{gFileIfaceHasUriScheme}
\alias{gFileIfaceIsNative}
\alias{gFileIfaceMakeDirectory}
\alias{gFileIfaceMakeSymbolicLink}
\alias{gFileIfaceMonitorDir}
\alias{gFileIfaceMonitorFile}
\alias{gFileIfaceMountEnclosingVolume}
\alias{gFileIfaceMountEnclosingVolumeFinish}
\alias{gFileIfaceMountMountable}
\alias{gFileIfaceMountMountableFinish}
\alias{gFileIfaceMove}
\alias{gFileIfaceOpenReadwrite}
\alias{gFileIfaceOpenReadwriteAsync}
\alias{gFileIfaceOpenReadwriteFinish}
\alias{gFileIfacePollMountable}
\alias{gFileIfacePollMountableFinish}
\alias{gFileIfacePrefixMatches}
\alias{gFileIfaceQueryFilesystemInfo}
\alias{gFileIfaceQueryFilesystemInfoAsync}
\alias{gFileIfaceQueryFilesystemInfoFinish}
\alias{gFileIfaceQueryInfo}
\alias{gFileIfaceQueryInfoAsync}
\alias{gFileIfaceQueryInfoFinish}
\alias{gFileIfaceQuerySettableAttributes}
\alias{gFileIfaceQueryWritableNamespaces}
\alias{gFileIfaceReadAsync}
\alias{gFileIfaceReadFinish}
\alias{gFileIfaceReadFn}
\alias{gFileIfaceReplace}
\alias{gFileIfaceReplaceAsync}
\alias{gFileIfaceReplaceFinish}
\alias{gFileIfaceReplaceReadwrite}
\alias{gFileIfaceReplaceReadwriteAsync}
\alias{gFileIfaceReplaceReadwriteFinish}
\alias{gFileIfaceResolveRelativePath}
\alias{gFileIfaceSetAttribute}
\alias{gFileIfaceSetAttributesAsync}
\alias{gFileIfaceSetAttributesFinish}
\alias{gFileIfaceSetAttributesFromInfo}
\alias{gFileIfaceSetDisplayName}
\alias{gFileIfaceSetDisplayNameAsync}
\alias{gFileIfaceSetDisplayNameFinish}
\alias{gFileIfaceStartMountable}
\alias{gFileIfaceStartMountableFinish}
\alias{gFileIfaceStopMountable}
\alias{gFileIfaceStopMountableFinish}
\alias{gFileIfaceTrash}
\alias{gFileIfaceUnmountMountable}
\alias{gFileIfaceUnmountMountableFinish}
\alias{gFileIfaceUnmountMountableWithOperation}
\alias{gFileIfaceUnmountMountableWithOperationFinish}
\alias{gFileInputStreamClassQueryInfo}
\alias{gFileInputStreamClassQueryInfoAsync}
\alias{gFileInputStreamClassQueryInfoFinish}
\alias{gFileMonitorClassCancel}
\alias{gFileOutputStreamClassGetEtag}
\alias{gFileOutputStreamClassQueryInfo}
\alias{gFileOutputStreamClassQueryInfoAsync}
\alias{gFileOutputStreamClassQueryInfoFinish}
\alias{gIOStreamClassCloseAsync}
\alias{gIOStreamClassCloseFinish}
\alias{gIOStreamClassCloseFn}
\alias{gIOStreamClassGetInputStream}
\alias{gIOStreamClassGetOutputStream}
\alias{gIconIfaceEqual}
\alias{gIconIfaceHash}
\alias{gInetAddressClassToBytes}
\alias{gInetAddressClassToString}
\alias{gInitableIfaceInit}
\alias{gInputStreamClassCloseAsync}
\alias{gInputStreamClassCloseFinish}
\alias{gInputStreamClassCloseFn}
\alias{gInputStreamClassReadFinish}
\alias{gInputStreamClassSkip}
\alias{gInputStreamClassSkipAsync}
\alias{gInputStreamClassSkipFinish}
\alias{gLoadableIconIfaceLoad}
\alias{gLoadableIconIfaceLoadAsync}
\alias{gLoadableIconIfaceLoadFinish}
\alias{gMountIfaceCanEject}
\alias{gMountIfaceCanUnmount}
\alias{gMountIfaceEject}
\alias{gMountIfaceEjectFinish}
\alias{gMountIfaceEjectWithOperation}
\alias{gMountIfaceEjectWithOperationFinish}
\alias{gMountIfaceGetDrive}
\alias{gMountIfaceGetIcon}
\alias{gMountIfaceGetName}
\alias{gMountIfaceGetRoot}
\alias{gMountIfaceGetUuid}
\alias{gMountIfaceGetVolume}
\alias{gMountIfaceGuessContentType}
\alias{gMountIfaceGuessContentTypeFinish}
\alias{gMountIfaceGuessContentTypeSync}
\alias{gMountIfaceRemount}
\alias{gMountIfaceRemountFinish}
\alias{gMountIfaceUnmount}
\alias{gMountIfaceUnmountFinish}
\alias{gMountIfaceUnmountWithOperation}
\alias{gMountIfaceUnmountWithOperationFinish}
\alias{gOutputStreamClassCloseAsync}
\alias{gOutputStreamClassCloseFinish}
\alias{gOutputStreamClassCloseFn}
\alias{gOutputStreamClassFlush}
\alias{gOutputStreamClassFlushAsync}
\alias{gOutputStreamClassFlushFinish}
\alias{gOutputStreamClassSplice}
\alias{gOutputStreamClassSpliceAsync}
\alias{gOutputStreamClassSpliceFinish}
\alias{gOutputStreamClassWriteAsync}
\alias{gOutputStreamClassWriteFinish}
\alias{gOutputStreamClassWriteFn}
\alias{gResolverClassLookupByAddress}
\alias{gResolverClassLookupByAddressAsync}
\alias{gResolverClassLookupByAddressFinish}
\alias{gResolverClassLookupByName}
\alias{gResolverClassLookupByNameAsync}
\alias{gResolverClassLookupByNameFinish}
\alias{gResolverClassLookupService}
\alias{gResolverClassLookupServiceAsync}
\alias{gResolverClassLookupServiceFinish}
\alias{gSeekableIfaceCanSeek}
\alias{gSeekableIfaceCanTruncate}
\alias{gSeekableIfaceSeek}
\alias{gSeekableIfaceTell}
\alias{gSeekableIfaceTruncateFn}
\alias{gSocketAddressClassGetFamily}
\alias{gSocketAddressClassGetNativeSize}
\alias{gSocketAddressClassToNative}
\alias{gSocketAddressEnumeratorClassNext}
\alias{gSocketAddressEnumeratorClassNextAsync}
\alias{gSocketAddressEnumeratorClassNextFinish}
\alias{gSocketConnectableIfaceEnumerate}
\alias{gSocketControlMessageClassGetLevel}
\alias{gSocketControlMessageClassGetSize}
\alias{gSocketControlMessageClassSerialize}
\alias{gSocketListenerClassChanged}
\alias{gVfsClassGetFileForPath}
\alias{gVfsClassGetFileForUri}
\alias{gVfsClassGetSupportedUriSchemes}
\alias{gVfsClassIsActive}
\alias{gVfsClassParseName}
\alias{gVolumeIfaceCanEject}
\alias{gVolumeIfaceCanMount}
\alias{gVolumeIfaceEject}
\alias{gVolumeIfaceEjectFinish}
\alias{gVolumeIfaceEjectWithOperation}
\alias{gVolumeIfaceEjectWithOperationFinish}
\alias{gVolumeIfaceEnumerateIdentifiers}
\alias{gVolumeIfaceGetActivationRoot}
\alias{gVolumeIfaceGetDrive}
\alias{gVolumeIfaceGetIcon}
\alias{gVolumeIfaceGetIdentifier}
\alias{gVolumeIfaceGetMount}
\alias{gVolumeIfaceGetName}
\alias{gVolumeIfaceGetUuid}
\alias{gVolumeIfaceMountFinish}
\alias{gVolumeIfaceMountFn}
\alias{gVolumeIfaceShouldAutomount}
\alias{gVolumeMonitorClassGetConnectedDrives}
\alias{gVolumeMonitorClassGetMountForUuid}
\alias{gVolumeMonitorClassGetMounts}
\alias{gVolumeMonitorClassGetVolumeForUuid}
\alias{gVolumeMonitorClassGetVolumes}
\alias{gtkActivatableIfaceSyncActionProperties}
\alias{gtkActivatableIfaceUpdate}
\alias{gtkToolShellIfaceGetIconSize}
\alias{gtkToolShellIfaceGetOrientation}
\alias{gtkToolShellIfaceGetReliefStyle}
\alias{gtkToolShellIfaceGetStyle}
\alias{gtkToolShellIfaceRebuildMenu}
\alias{pangoRendererClassDrawGlyphItem}
\name{virtuals}
\title{Virtual Methods}
\description{This function is a wrapper for a virtual method in a \code{GObject} 
  class structure. The wrapper is normally only useful for calling virtuals in a 
  parent class inside implementations of new \code{GObject} classes.
}
\author{Michael Lawrence}
\seealso{gClass}
\keyword{internal}
\keyword{interface}
