\alias{atkActionDoAction}
\name{atkActionDoAction}
\title{atkActionDoAction}
\description{Perform the specified action on the object.}
\usage{atkActionDoAction(object, i)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkAction}}]  a \code{\link{GObject}} instance that implements AtkActionIface}
\item{\verb{i}}{[integer]  the action index corresponding to the action to be performed }
}
\value{[logical] \code{TRUE} if success, \code{FALSE} otherwise}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
