\alias{atkComponentGetLayer}
\name{atkComponentGetLayer}
\title{atkComponentGetLayer}
\description{Gets the layer of the component.}
\usage{atkComponentGetLayer(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkComponent}}]  an \code{\link{AtkComponent}}}}
\value{[\code{\link{AtkLayer}}]  an \code{\link{AtkLayer}} which is the layer of the component}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
