\alias{atkComponentGrabFocus}
\name{atkComponentGrabFocus}
\title{atkComponentGrabFocus}
\description{Grabs focus for this \code{component}.}
\usage{atkComponentGrabFocus(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkComponent}}]  an \code{\link{AtkComponent}}}}
\value{[logical] \code{TRUE} if successful, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
