\alias{atkComponentSetSize}
\name{atkComponentSetSize}
\title{atkComponentSetSize}
\description{Set the size of the \code{component} in terms of width and height.}
\usage{atkComponentSetSize(object, width, height)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkComponent}}]  an \code{\link{AtkComponent}}}
\item{\verb{width}}{[integer]  width to set for \code{component}}
\item{\verb{height}}{[integer]  height to set for \code{component}}
}
\value{[logical] \code{TRUE} or \code{FALSE} whether the size was set or not}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
