\alias{atkDocumentGetDocumentType}
\name{atkDocumentGetDocumentType}
\title{atkDocumentGetDocumentType}
\description{Gets a string indicating the document type.}
\usage{atkDocumentGetDocumentType(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkDocument}}]  a \code{\link{GObject}} instance that implements AtkDocumentIface}}
\value{[character]  a string indicating the document type}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
