\alias{atkEditableTextCopyText}
\name{atkEditableTextCopyText}
\title{atkEditableTextCopyText}
\description{Copy text from \code{start.pos} up to, but not including \code{end.pos} 
to the clipboard.}
\usage{atkEditableTextCopyText(object, start.pos, end.pos)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkEditableText}}]  an \code{\link{AtkEditableText}}}
\item{\verb{start.pos}}{[integer]  start position}
\item{\verb{end.pos}}{[integer]  end position}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
