\alias{atkEditableTextDeleteText}
\name{atkEditableTextDeleteText}
\title{atkEditableTextDeleteText}
\description{Delete text \code{start.pos} up to, but not including \code{end.pos}.}
\usage{atkEditableTextDeleteText(object, start.pos, end.pos)}
\arguments{
\item{\verb{object}}{[\verb{\link{AtkEditableText}}]  an \verb{\link{AtkEditableText}}}
\item{\verb{start.pos}}{[integer]  start position}
\item{\verb{end.pos}}{[integer]  end position}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
