\alias{atkEditableTextPasteText}
\name{atkEditableTextPasteText}
\title{atkEditableTextPasteText}
\description{Paste text from clipboard to specified \code{position}.}
\usage{atkEditableTextPasteText(object, position)}
\arguments{
\item{\code{object}}{[\code{\link{AtkEditableText}}]  an \code{\link{AtkEditableText}}}
\item{\code{position}}{[integer]  position to paste}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
