\alias{atkEditableTextSetTextContents}
\name{atkEditableTextSetTextContents}
\title{atkEditableTextSetTextContents}
\description{Set text contents of \code{text}.}
\usage{atkEditableTextSetTextContents(object, string)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkEditableText}}]  an \code{\link{AtkEditableText}}}
\item{\verb{string}}{[character]  string to set for text contents of \code{text}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
