\alias{atkGObjectAccessibleForObject}
\name{atkGObjectAccessibleForObject}
\title{atkGObjectAccessibleForObject}
\description{Gets the accessible object for the specified \code{obj}.}
\usage{atkGObjectAccessibleForObject(obj)}
\arguments{\item{\verb{obj}}{[\code{\link{GObject}}]  a \code{\link{GObject}}}}
\value{[\code{\link{AtkObject}}]  a \code{\link{AtkObject}} which is the accessible object for the \code{obj}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
