\alias{atkGObjectAccessibleGetObject}
\name{atkGObjectAccessibleGetObject}
\title{atkGObjectAccessibleGetObject}
\description{Gets the GObject for which \code{obj} is the accessible object.}
\usage{atkGObjectAccessibleGetObject(object)}
\arguments{\item{\verb{object}}{[\verb{\link{AtkGObjectAccessible}}]  a \verb{\link{AtkObject}}}}
\value{[\verb{\link{GObject}}]  a \verb{\link{GObject}} which is the object for which \code{obj} is the accessible objedct}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
