\alias{atkGObjectAccessibleGetObject}
\name{atkGObjectAccessibleGetObject}
\title{atkGObjectAccessibleGetObject}
\description{Gets the GObject for which \code{obj} is the accessible object.}
\usage{atkGObjectAccessibleGetObject(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkGObjectAccessible}}]  a \code{\link{AtkGObjectAccessible}}}}
\value{[\code{\link{GObject}}]  a \code{\link{GObject}} which is the object for which \code{obj} is the accessible object}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
