\alias{atkHyperlinkGetNAnchors}
\name{atkHyperlinkGetNAnchors}
\title{atkHyperlinkGetNAnchors}
\description{Gets the number of anchors associated with this hyperlink.}
\usage{atkHyperlinkGetNAnchors(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkHyperlink}}]  an \code{\link{AtkHyperlink}}}}
\value{[integer]  the number of anchors associated with this hyperlink}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
