\alias{atkHyperlinkGetStartIndex}
\name{atkHyperlinkGetStartIndex}
\title{atkHyperlinkGetStartIndex}
\description{Gets the index with the hypertext document at which this link begins.}
\usage{atkHyperlinkGetStartIndex(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkHyperlink}}]  an \code{\link{AtkHyperlink}}}}
\value{[integer]  the index with the hypertext document at which this link begins}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
