\alias{atkHyperlinkIsSelectedLink}
\name{atkHyperlinkIsSelectedLink}
\title{atkHyperlinkIsSelectedLink}
\description{Determines whether this AtkHyperlink is selected}
\usage{atkHyperlinkIsSelectedLink(object)}
\arguments{\item{\code{object}}{[\code{\link{AtkHyperlink}}]  an \code{\link{AtkHyperlink}}}}
\details{  Since  ATK 1.4

@Deprecated: This method is deprecated since ATK version 1.8.
Please use ATK\_STATE\_SELECTED to indicate when a hyperlink within a
Hypertext container is selected.}
\value{[logical]  True is the AtkHyperlink is selected, False otherwise}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
