\alias{atkImageGetImageDescription}
\name{atkImageGetImageDescription}
\title{atkImageGetImageDescription}
\description{Get a textual description of this image.}
\usage{atkImageGetImageDescription(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkImage}}]  a \code{\link{GObject}} instance that implements AtkImageIface}}
\value{[character]  a string representing the image description}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
