\alias{atkImageSetImageDescription}
\name{atkImageSetImageDescription}
\title{atkImageSetImageDescription}
\description{Sets the textual description for this image.}
\usage{atkImageSetImageDescription(object, description)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkImage}}]  a \code{\link{GObject}} instance that implements AtkImageIface}
\item{\verb{description}}{[character]  a string description to set for \code{image}}
}
\value{[logical]  boolean TRUE, or FALSE if operation could
not be completed.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
