\alias{atkImplementorRefAccessible}
\name{atkImplementorRefAccessible}
\title{atkImplementorRefAccessible}
\description{Gets a reference to an object's \code{\link{AtkObject}} implementation, if
the object implements \verb{AtkObjectIface}}
\usage{atkImplementorRefAccessible(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkImplementor}}]  The \code{\link{GObject}} instance which should implement \verb{AtkImplementorIface}
if a non-null return value is required.}}
\value{[\code{\link{AtkObject}}]  a reference to an object's \code{\link{AtkObject}} implementation}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
