\alias{atkObjectFactoryCreateAccessible}
\name{atkObjectFactoryCreateAccessible}
\title{atkObjectFactoryCreateAccessible}
\description{Provides an \code{\link{AtkObject}} that implements an accessibility interface 
on behalf of \code{obj}}
\usage{atkObjectFactoryCreateAccessible(object, obj)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkObjectFactory}}]  The \code{\link{AtkObjectFactory}} associated with \code{obj}'s
object type}
\item{\verb{obj}}{[\code{\link{GObject}}]  a \code{\link{GObject}} }
}
\value{[\code{\link{AtkObject}}]  an \code{\link{AtkObject}} that implements an accessibility interface
on behalf of \code{obj}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
