\alias{atkObjectFactoryGetAccessibleType}
\name{atkObjectFactoryGetAccessibleType}
\title{atkObjectFactoryGetAccessibleType}
\description{Gets the GType of the accessible which is created by the factory.}
\usage{atkObjectFactoryGetAccessibleType(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkObjectFactory}}]  an \code{\link{AtkObjectFactory}} }}
\value{[\code{\link{GType}}]  the type of the accessible which is created by the \code{factory}.
The value G_TYPE_INVALID is returned if no type if found.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
