\alias{atkObjectGetLayer}
\name{atkObjectGetLayer}
\title{atkObjectGetLayer}
\description{
Gets the layer of the accessible.
\strong{WARNING: \code{atk_object_get_layer} is deprecated and should not be used in newly-written code. Use atk\_component\_get\_layer instead.}
}
\usage{atkObjectGetLayer(object)}
\arguments{\item{\code{object}}{[\code{\link{AtkObject}}]  an \code{\link{AtkObject}}}}
\details{Returns:  }
\value{[\code{\link{AtkLayer}}]  an \code{\link{AtkLayer}} which is the layer of the accessible}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
