\alias{atkObjectGetRole}
\name{atkObjectGetRole}
\title{atkObjectGetRole}
\description{Gets the role of the accessible.}
\usage{atkObjectGetRole(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkObject}}]  an \code{\link{AtkObject}}}}
\value{[\code{\link{AtkRole}}]  an \code{\link{AtkRole}} which is the role of the accessible}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
