\alias{atkObjectNotifyStateChange}
\name{atkObjectNotifyStateChange}
\title{atkObjectNotifyStateChange}
\description{Emits a state-change signal for the specified state.}
\usage{atkObjectNotifyStateChange(object, state, value)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkObject}}]  an \code{\link{AtkObject}}}
\item{\verb{state}}{[numeric]  an \verb{numeric} whose state is changed}
\item{\verb{value}}{[logical]  a gboolean which indicates whether the state is being set on or off}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
