\alias{atkObjectRefRelationSet}
\name{atkObjectRefRelationSet}
\title{atkObjectRefRelationSet}
\description{Gets the \code{\link{AtkRelationSet}} associated with the object.}
\usage{atkObjectRefRelationSet(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkObject}}]  an \code{\link{AtkObject}}}}
\value{[\code{\link{AtkRelationSet}}]  an \code{\link{AtkRelationSet}} representing the relation set of the object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
