\alias{atkObjectSetDescription}
\name{atkObjectSetDescription}
\title{atkObjectSetDescription}
\description{Sets the accessible description of the accessible.}
\usage{atkObjectSetDescription(object, description)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkObject}}]  an \code{\link{AtkObject}}}
\item{\verb{description}}{[character]  a character string to be set as the accessible description}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
