\alias{atkObjectSetName}
\name{atkObjectSetName}
\title{atkObjectSetName}
\description{Sets the accessible name of the accessible.}
\usage{atkObjectSetName(object, name)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkObject}}]  an \code{\link{AtkObject}}}
\item{\verb{name}}{[character]  a character string to be set as the accessible name}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
