\alias{atkObjectSetParent}
\name{atkObjectSetParent}
\title{atkObjectSetParent}
\description{Sets the accessible parent of the accessible.}
\usage{atkObjectSetParent(object, parent)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkObject}}]  an \code{\link{AtkObject}}}
\item{\verb{parent}}{[\code{\link{AtkObject}}]  an \code{\link{AtkObject}} to be set as the accessible parent}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
