\alias{atkObjectSetRole}
\name{atkObjectSetRole}
\title{atkObjectSetRole}
\description{Sets the role of the accessible.}
\usage{atkObjectSetRole(object, role)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkObject}}]  an \code{\link{AtkObject}}}
\item{\verb{role}}{[\code{\link{AtkRole}}]  an \code{\link{AtkRole}} to be set as the role}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
