\alias{atkRelationAddTarget}
\name{atkRelationAddTarget}
\title{atkRelationAddTarget}
\description{Adds the specified AtkObject to the target for the relation, if it is
not already present.}
\usage{atkRelationAddTarget(object, target)}
\arguments{
\item{\code{object}}{[\code{\link{AtkRelation}}]  an \code{\link{AtkRelation}}}
\item{\code{target}}{[\code{\link{AtkObject}}]  an \code{\link{AtkObject}}}
}
\details{  Since  1.9}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
