\alias{atkRelationSetGetRelation}
\name{atkRelationSetGetRelation}
\title{atkRelationSetGetRelation}
\description{Determines the relation at the specified position in the relation set.}
\usage{atkRelationSetGetRelation(object, i)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkRelationSet}}]  an \code{\link{AtkRelationSet}}}
\item{\verb{i}}{[integer]  a gint representing a position in the set, starting from 0.}
}
\value{[\code{\link{AtkRelation}}]  a \code{\link{AtkRelation}}, which is the relation at position i in the set.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
