\alias{atkRemoveFocusTracker}
\name{atkRemoveFocusTracker}
\title{atkRemoveFocusTracker}
\description{Removes the specified focus tracker from the list of functions
to be called when any object receives focus.}
\usage{atkRemoveFocusTracker(tracker.id)}
\arguments{\item{\verb{tracker.id}}{[numeric]  the id of the focus tracker to remove}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
