\alias{atkRemoveKeyEventListener}
\name{atkRemoveKeyEventListener}
\title{atkRemoveKeyEventListener}
\description{Removes the specified event listener}
\usage{atkRemoveKeyEventListener(listener.id)}
\arguments{\item{\verb{listener.id}}{[numeric]  the id of the event listener to remove}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
