\alias{atkRoleRegister}
\name{atkRoleRegister}
\title{atkRoleRegister}
\description{Registers the role specified by \code{name}.}
\usage{atkRoleRegister(name)}
\arguments{\item{\verb{name}}{[character]  a character string describing the new role.}}
\value{[\code{\link{AtkRole}}]  an \code{\link{AtkRole}} for the new role.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
