\alias{atkSelectionAddSelection}
\name{atkSelectionAddSelection}
\title{atkSelectionAddSelection}
\description{Adds the specified accessible child of the object to the
object's selection.}
\usage{atkSelectionAddSelection(object, i)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkSelection}}]  a \code{\link{GObject}} instance that implements AtkSelectionIface}
\item{\verb{i}}{[integer]  a \verb{integer} specifying the child index.}
}
\value{[logical]  TRUE if success, FALSE otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
