\alias{atkSelectionClearSelection}
\name{atkSelectionClearSelection}
\title{atkSelectionClearSelection}
\description{Clears the selection in the object so that no children in the object
are selected.}
\usage{atkSelectionClearSelection(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkSelection}}]  a \code{\link{GObject}} instance that implements AtkSelectionIface}}
\value{[logical]  TRUE if success, FALSE otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
