\alias{atkStateSetAddStates}
\name{atkStateSetAddStates}
\title{atkStateSetAddStates}
\description{Add the states for the specified types to the current state set.}
\usage{atkStateSetAddStates(object, types)}
\arguments{
\item{\verb{object}}{[\verb{\link{AtkStateSet}}]  an \verb{\link{AtkStateSet}}}
\item{\verb{types}}{[\code{\link{AtkStateType}}]  a list of \code{\link{AtkStateType}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
