\alias{atkStateSetAndSets}
\name{atkStateSetAndSets}
\title{atkStateSetAndSets}
\description{Constructs the intersection of the two sets, returning \code{NULL} if the
intersection is empty.}
\usage{atkStateSetAndSets(object, compare.set)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkStateSet}}]  an \code{\link{AtkStateSet}}}
\item{\verb{compare.set}}{[\code{\link{AtkStateSet}}]  another \code{\link{AtkStateSet}}}
}
\value{[\code{\link{AtkStateSet}}]  a new \code{\link{AtkStateSet}} which is the intersection of the two sets.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
