\alias{atkStateSetClearStates}
\name{atkStateSetClearStates}
\title{atkStateSetClearStates}
\description{Removes all states from the state set.}
\usage{atkStateSetClearStates(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkStateSet}}]  an \code{\link{AtkStateSet}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
