\alias{atkStateSetOrSets}
\name{atkStateSetOrSets}
\title{atkStateSetOrSets}
\description{Constructs the union of the two sets.}
\usage{atkStateSetOrSets(object, compare.set)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkStateSet}}]  an \code{\link{AtkStateSet}}}
\item{\verb{compare.set}}{[\code{\link{AtkStateSet}}]  another \code{\link{AtkStateSet}}}
}
\value{[\code{\link{AtkStateSet}}]  a new \code{\link{AtkStateSet}} which is the union of the two sets,
returning \code{NULL} is empty.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
