\alias{AtkStreamableContent}
\name{AtkStreamableContent}
\title{AtkStreamableContent}
\description{The ATK interface which provides access to streamable content.}
\section{Methods and Functions}{
\link{atkStreamableContentGetNMimeTypes}(object)
\link{atkStreamableContentGetMimeType}(object, i)
\link{atkStreamableContentGetStream}(object, mime.type)
}
\section{Hierarchy}{\preformatted{  GInterface
   +----AtkStreamableContent}}
\section{Structures}{\describe{\item{\code{AtkStreamableContent}}{
The AtkStreamableContent structure does not contain any fields.

}}}
\references{\url{http://developer.gnome.org/doc/API/2.0/atk/AtkStreamableContent.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
