\alias{atkTableSetCaption}
\name{atkTableSetCaption}
\title{atkTableSetCaption}
\description{Sets the caption for the table.}
\usage{atkTableSetCaption(object, caption)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkTable}}]  a GObject instance that implements AtkTableIface}
\item{\verb{caption}}{[\code{\link{AtkObject}}]  a \code{\link{AtkObject}} representing the caption to set for \code{table}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
