\alias{atkTableSetColumnHeader}
\name{atkTableSetColumnHeader}
\title{atkTableSetColumnHeader}
\description{Sets the specified column header to \code{header}.}
\usage{atkTableSetColumnHeader(object, column, header)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkTable}}]  a GObject instance that implements AtkTableIface}
\item{\verb{column}}{[integer]  a \verb{integer} representing a column in \code{table}}
\item{\verb{header}}{[\code{\link{AtkObject}}]  an \code{\link{AtkTable}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
