\alias{atkTableSetRowDescription}
\name{atkTableSetRowDescription}
\title{atkTableSetRowDescription}
\description{Sets the description text for the specified \code{row} of \code{table}.}
\usage{atkTableSetRowDescription(object, row, description)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkTable}}]  a GObject instance that implements AtkTableIface}
\item{\verb{row}}{[integer]  a \verb{integer} representing a row in \code{table}}
\item{\verb{description}}{[character]  a \verb{character} representing the description text
to set for the specified \code{row} of \code{table}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
