\alias{atkTableSetRowHeader}
\name{atkTableSetRowHeader}
\title{atkTableSetRowHeader}
\description{Sets the specified row header to \code{header}.}
\usage{atkTableSetRowHeader(object, row, header)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkTable}}]  a GObject instance that implements AtkTableIface}
\item{\verb{row}}{[integer]  a \verb{integer} representing a row in \code{table}}
\item{\verb{header}}{[\code{\link{AtkObject}}]  an \code{\link{AtkTable}} }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
