\alias{atkTableSetSummary}
\name{atkTableSetSummary}
\title{atkTableSetSummary}
\description{Sets the summary description of the table.}
\usage{atkTableSetSummary(object, accessible)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkTable}}]  a GObject instance that implements AtkTableIface}
\item{\verb{accessible}}{[\code{\link{AtkObject}}]  an \code{\link{AtkObject}} representing the summary description
to set for \code{table}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
