\alias{atkTextAttributeGetName}
\name{atkTextAttributeGetName}
\title{atkTextAttributeGetName}
\description{Gets the name corresponding to the \code{\link{AtkTextAttribute}}}
\usage{atkTextAttributeGetName(attr)}
\arguments{\item{\verb{attr}}{[\code{\link{AtkTextAttribute}}]  The \code{\link{AtkTextAttribute}} whose name is required}}
\value{[character]  a string containing the name; this string should not be freed}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
