\alias{atkTextAttributeGetValue}
\name{atkTextAttributeGetValue}
\title{atkTextAttributeGetValue}
\description{Gets the value for the index of the \code{\link{AtkTextAttribute}}}
\usage{atkTextAttributeGetValue(attr, index)}
\arguments{
\item{\code{attr}}{[\code{\link{AtkTextAttribute}}]  The \code{\link{AtkTextAttribute}} for which a value is required}
\item{\code{index}}{[integer]  The index of the required value}
}
\value{[character]  .}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
