\alias{atkTextGetCharacterAtOffset}
\name{atkTextGetCharacterAtOffset}
\title{atkTextGetCharacterAtOffset}
\description{Gets the specified text.}
\usage{atkTextGetCharacterAtOffset(object, offset)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkText}}]  an \code{\link{AtkText}}}
\item{\verb{offset}}{[integer]  position}
}
\value{[numeric]  the character at \code{offset}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
