\alias{atkTextGetText}
\name{atkTextGetText}
\title{atkTextGetText}
\description{Gets the specified text.}
\usage{atkTextGetText(object, start.offset, end.offset)}
\arguments{
\item{\verb{object}}{[\code{\link{AtkText}}]  an \code{\link{AtkText}}}
\item{\verb{start.offset}}{[integer]  start position}
\item{\verb{end.offset}}{[integer]  end position}
}
\value{[character]  the text from \code{start.offset} up to, but not including \code{end.offset}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
