\alias{atkValueGetCurrentValue}
\name{atkValueGetCurrentValue}
\title{atkValueGetCurrentValue}
\description{Gets the value of this object.}
\usage{atkValueGetCurrentValue(object)}
\arguments{\item{\verb{object}}{[\code{\link{AtkValue}}]  a GObject instance that implements AtkValueIface}}
\value{
A list containing the following elements:
\item{\verb{value}}{[R object]  a \verb{R object} representing the current accessible value}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
