\alias{cairo-PNG-Support}
\name{cairo-PNG-Support}
\title{PNG Support}
\description{Reading and writing PNG images}
\section{Methods and Functions}{
\link{cairoImageSurfaceCreateFromPng}(filename)
\link{cairoImageSurfaceCreateFromPngStream}(con)
\link{cairoSurfaceWriteToPng}(surface, filename)
\link{cairoSurfaceWriteToPngStream}(surface, con)
}
\references{\url{http://developer.gnome.org/doc/API/2.0/cairo/cairo-PNG-Support.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
