\alias{cairo-PNG-Support}
\name{cairo-PNG-Support}
\title{PNG Support}
\description{Reading and writing PNG images}
\section{Methods and Functions}{
\code{\link{cairoImageSurfaceCreateFromPng}(filename)}\cr
\code{\link{cairoImageSurfaceCreateFromPngStream}(con)}\cr
\code{\link{cairoSurfaceWriteToPng}(surface, filename)}\cr
\code{\link{cairoSurfaceWriteToPngStream}(surface, con)}\cr
}
\section{Detailed Description}{The PNG functions allow reading PNG images into image surfaces, and writing
any surface to a PNG file.}
\references{\url{http://www.cairographics.org/manual/cairo-PNG-Support.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
