\alias{cairo-Transformations}
\name{cairo-Transformations}
\title{Transformations}
\description{Manipulating the current transformation matrix}
\section{Methods and Functions}{
\link{cairoTranslate}(cr, tx, ty)
\link{cairoScale}(cr, sx, sy)
\link{cairoRotate}(cr, angle)
\link{cairoTransform}(cr, matrix)
\link{cairoSetMatrix}(cr, matrix)
\link{cairoGetMatrix}(cr, matrix)
\link{cairoIdentityMatrix}(cr)
\link{cairoUserToDevice}(cr, x, y)
\link{cairoUserToDeviceDistance}(cr, dx, dy)
\link{cairoDeviceToUser}(cr, x, y)
\link{cairoDeviceToUserDistance}(cr, dx, dy)
}
\references{\url{http://developer.gnome.org/doc/API/2.0/cairo/cairo-Transformations.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
