\alias{cairo-Version-Information}
\name{cairo-Version-Information}
\title{Version Information}
\description{Compile and run time version checks}
\section{Methods and Functions}{\code{\link{cairoVersionString}()}\cr}
\references{\url{http://www.cairographics.org/manual/cairo-Version-Information.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
