\alias{CAIRO}
\name{CAIRO}
\title{CAIRO}
\description{Cairo is a 2D graphics library with support for multiple output devices. Currently supported output targets include the X Window System, win32, and image buffers.}
\details{
The RGtk binding to the CAIRO library consists of the following components:
\describe{
\item{\link{cairo-font-face}}{Base class for font faces}
\item{\link{cairo-font-options}}{How a font should be rendered}
\item{\link{cairo-image-surface}}{Rendering to memory buffers}
\item{\link{cairo-matrix}}{Generic matrix operations}
\item{\link{cairo-paths}}{Creating paths and manipulating path data}
\item{\link{cairo-pattern}}{Sources for drawing}
\item{\link{cairo-pdf-surface}}{Rendering PDF documents}
\item{\link{cairo-png-functions}}{Reading and writing PNG images}
\item{\link{cairo-ps-surface}}{Rendering PostScript documents}
\item{\link{cairo-scaled-font}}{Font face at particular size and options}
\item{\link{cairo-error-status}}{Decoding cairo's status}
\item{\link{cairo-surface}}{Base class for surfaces}
\item{\link{cairo-svg-surface}}{Rendering SVG documents}
\item{\link{cairo-text}}{Rendering text and glyphs}
\item{\link{cairo-transformations}}{Manipulating the current transformation matrix}
\item{\link{cairo-types}}{Generic data types}
\item{\link{cairo-user-font}}{Font support with font data provided by the user}
\item{\link{cairo-version-info}}{Compile-time and run-time version checks.}
\item{\link{cairo-context}}{The cairo drawing context}
}
}
\references{\url{https://www.cairographics.org/manual/}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{interface}
