\alias{cairoFillExtents}
\name{cairoFillExtents}
\title{cairoFillExtents}
\description{\emph{undocumented
}}
\usage{cairoFillExtents(cr)}
\arguments{\item{\code{cr}}{[\code{\link{cairo_t}}] }}
\value{
A list containing the following elements:
\item{\code{x1}}{[numeric] }
\item{\code{y1}}{[numeric] }
\item{\code{x2}}{[numeric] }
\item{\code{y2}}{[numeric] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
