\alias{cairoFontExtents}
\name{cairoFontExtents}
\title{cairoFontExtents}
\description{Gets the font extents for the currently selected font.}
\usage{cairoFontExtents(cr)}
\arguments{\item{\code{cr}}{[\code{\link{cairo_t}}]  a \code{\link{cairo_t}}}}
\value{
A list containing the following elements:
\item{\code{extents}}{[\code{\link{cairo_font_extents_t}}]  a \code{\link{cairo_font_extents_t}} object into which the results
will be stored.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
