\alias{cairoFontExtents}
\name{cairoFontExtents}
\title{cairoFontExtents}
\description{Gets the font extents for the currently selected font.}
\usage{cairoFontExtents(cr)}
\arguments{\item{\verb{cr}}{[\code{\link{Cairo}}]  a \code{\link{Cairo}}}}
\value{
A list containing the following elements:
\item{\verb{extents}}{[\code{\link{CairoFontExtents}}]  a \code{\link{CairoFontExtents}} object into which the results
will be stored.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
