\alias{cairoFontFaceGetType}
\name{cairoFontFaceGetType}
\title{cairoFontFaceGetType}
\description{This function returns the type of the backend used to create
a font face. See \code{\link{CairoFontType}} for available types.}
\usage{cairoFontFaceGetType(font.face)}
\arguments{\item{\verb{font.face}}{[\code{\link{CairoFontFace}}]  a font face}}
\details{  Since 1.2}
\value{[\code{\link{CairoFontType}}]  The type of \code{font.face}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
