\alias{cairoFontFaceStatus}
\name{cairoFontFaceStatus}
\title{cairoFontFaceStatus}
\description{Checks whether an error has previously occurred for this
font face}
\usage{cairoFontFaceStatus(font.face)}
\arguments{\item{\code{font.face}}{[\code{\link{cairo_font_face_t}}]  a \code{\link{cairo_font_face_t}}}}
\value{[\code{\link{cairo_status_t}}]  \code{CAIRO_STATUS_SUCCESS} or another error such as
  \code{CAIRO_STATUS_NO_MEMORY}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
