\alias{cairoFontOptionsEqual}
\name{cairoFontOptionsEqual}
\title{cairoFontOptionsEqual}
\description{Compares two font options objects for equality.}
\usage{cairoFontOptionsEqual(options, other)}
\arguments{
\item{\code{options}}{[\code{\link{cairo_font_options_t}}]  a \code{\link{cairo_font_options_t}}}
\item{\code{other}}{[\code{\link{cairo_font_options_t}}]  another \code{\link{cairo_font_options_t}}}
}
\value{[cairo_bool_t]  \code{TRUE} if all fields of the two font options objects match}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
