\alias{cairoFontOptionsGetAntialias}
\name{cairoFontOptionsGetAntialias}
\title{cairoFontOptionsGetAntialias}
\description{Gets the antialiasing mode for the font options object.}
\usage{cairoFontOptionsGetAntialias(options)}
\arguments{\item{\verb{options}}{[\verb{\link{CairoFontOptions}}]  a \verb{\link{CairoFontOptions}}}}
\value{[\verb{\link{CairoAntialias}}]  the antialiasing mode}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
