\alias{cairoFontOptionsGetHintMetrics}
\name{cairoFontOptionsGetHintMetrics}
\title{cairoFontOptionsGetHintMetrics}
\description{Gets the metrics hinting mode for the font options object.
See the documentation for \code{\link{cairo_hint_metrics_t}} for full details.}
\usage{cairoFontOptionsGetHintMetrics(options)}
\arguments{\item{\code{options}}{[\code{\link{cairo_font_options_t}}]  a \code{\link{cairo_font_options_t}}}}
\value{[\code{\link{cairo_hint_metrics_t}}]  the metrics hinting mode for the font options object}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
