\alias{cairoFontOptionsGetHintStyle}
\name{cairoFontOptionsGetHintStyle}
\title{cairoFontOptionsGetHintStyle}
\description{Gets the hint style for font outlines for the font options object.
See the documentation for \code{\link{CairoHintStyle}} for full details.}
\usage{cairoFontOptionsGetHintStyle(options)}
\arguments{\item{\verb{options}}{[\code{\link{CairoFontOptions}}]  a \code{\link{CairoFontOptions}}}}
\value{[\code{\link{CairoHintStyle}}]  the hint style for the font options object}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
