\alias{cairoFontOptionsGetSubpixelOrder}
\name{cairoFontOptionsGetSubpixelOrder}
\title{cairoFontOptionsGetSubpixelOrder}
\description{Gets the subpixel order for the font options object.
See the documentation for \verb{\link{CairoSubpixelOrder}} for full details.}
\usage{cairoFontOptionsGetSubpixelOrder(options)}
\arguments{\item{\verb{options}}{[\verb{\link{CairoFontOptions}}]  a \verb{\link{CairoFontOptions}}}}
\value{[\verb{\link{CairoSubpixelOrder}}]  the subpixel order for the font options object}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
