\alias{cairoFontOptionsSetAntialias}
\name{cairoFontOptionsSetAntialias}
\title{cairoFontOptionsSetAntialias}
\description{Sets the antialiasing mode for the font options object. This
specifies the type of antialiasing to do when rendering text.}
\usage{cairoFontOptionsSetAntialias(options, antialias)}
\arguments{
\item{\verb{options}}{[\code{\link{CairoFontOptions}}]  a \code{\link{CairoFontOptions}}}
\item{\verb{antialias}}{[\code{\link{CairoAntialias}}]  the new antialiasing mode}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
