\alias{cairoFontOptionsSetHintMetrics}
\name{cairoFontOptionsSetHintMetrics}
\title{cairoFontOptionsSetHintMetrics}
\description{Sets the metrics hinting mode for the font options object. This
controls whether metrics are quantized to integer values in
device units.
See the documentation for \code{\link{cairo_hint_metrics_t}} for full details.}
\usage{cairoFontOptionsSetHintMetrics(options, hint.metrics)}
\arguments{
\item{\code{options}}{[\code{\link{cairo_font_options_t}}]  a \code{\link{cairo_font_options_t}}}
\item{\code{hint.metrics}}{[\code{\link{cairo_hint_metrics_t}}]  the new metrics hinting mode}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
