\alias{cairoFontOptionsSetHintStyle}
\name{cairoFontOptionsSetHintStyle}
\title{cairoFontOptionsSetHintStyle}
\description{Sets the hint style for font outlines for the font options object.
This controls whether to fit font outlines to the pixel grid,
and if so, whether to optimize for fidelity or contrast.
See the documentation for \code{\link{CairoHintStyle}} for full details.}
\usage{cairoFontOptionsSetHintStyle(options, hint.style)}
\arguments{
\item{\verb{options}}{[\code{\link{CairoFontOptions}}]  a \code{\link{CairoFontOptions}}}
\item{\verb{hint.style}}{[\code{\link{CairoHintStyle}}]  the new hint style}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
