\alias{cairoFontOptionsStatus}
\name{cairoFontOptionsStatus}
\title{cairoFontOptionsStatus}
\description{Checks whether an error has previously occurred for this
font options object}
\usage{cairoFontOptionsStatus(options)}
\arguments{\item{\code{options}}{[\code{\link{cairo_font_options_t}}]  a \code{\link{cairo_font_options_t}}}}
\value{[\code{\link{cairo_status_t}}]  \code{CAIRO_STATUS_SUCCESS} or \code{CAIRO_STATUS_NO_MEMORY}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
