\alias{cairoGetAntialias}
\name{cairoGetAntialias}
\title{cairoGetAntialias}
\description{Gets the current shape antialiasing mode, as set by \code{cairoSetShapeAntialias()}.}
\usage{cairoGetAntialias(cr)}
\arguments{\item{\code{cr}}{[\code{\link{cairo_t}}]  a cairo context}}
\value{[\code{\link{cairo_antialias_t}}]  the current shape antialiasing mode.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
