\alias{cairoGetAntialias}
\name{cairoGetAntialias}
\title{cairoGetAntialias}
\description{Gets the current shape antialiasing mode, as set by \code{cairoSetShapeAntialias()}.}
\usage{cairoGetAntialias(cr)}
\arguments{\item{\verb{cr}}{[\verb{\link{Cairo}}]  a cairo context}}
\value{[\verb{\link{CairoAntialias}}]  the current shape antialiasing mode.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
