\alias{cairoGetAntialias}
\name{cairoGetAntialias}
\title{cairoGetAntialias}
\description{Gets the current shape antialiasing mode, as set by \code{\link{cairoSetAntialias}}.}
\usage{cairoGetAntialias(cr)}
\arguments{\item{\verb{cr}}{[\code{\link{Cairo}}]  a cairo context}}
\value{[\code{\link{CairoAntialias}}]  the current shape antialiasing mode.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
