\alias{cairoGetFillRule}
\name{cairoGetFillRule}
\title{cairoGetFillRule}
\description{Gets the current fill rule, as set by \code{\link{cairoSetFillRule}}.}
\usage{cairoGetFillRule(cr)}
\arguments{\item{\verb{cr}}{[\code{\link{Cairo}}]  a cairo context}}
\value{[\code{\link{CairoFillRule}}]  the current fill rule.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
