\alias{cairoGetFontMatrix}
\name{cairoGetFontMatrix}
\title{cairoGetFontMatrix}
\description{Stores the current font matrix into \code{matrix}. See
\code{\link{cairoSetFontMatrix}}.}
\usage{cairoGetFontMatrix(cr, matrix)}
\arguments{
\item{\verb{cr}}{[\verb{\link{Cairo}}]  a \verb{\link{Cairo}}}
\item{\verb{matrix}}{[\verb{\link{CairoMatrix}}]  return value for the matrix}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
