\alias{cairoGetLineCap}
\name{cairoGetLineCap}
\title{cairoGetLineCap}
\description{Gets the current line cap style, as set by \code{\link{cairoSetLineCap}}.}
\usage{cairoGetLineCap(cr)}
\arguments{\item{\verb{cr}}{[\code{\link{Cairo}}]  a cairo context}}
\value{[\code{\link{CairoLineCap}}]  the current line cap style.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
