\alias{cairoGetLineJoin}
\name{cairoGetLineJoin}
\title{cairoGetLineJoin}
\description{Gets the current line join style, as set by \code{\link{cairoSetLineJoin}}.}
\usage{cairoGetLineJoin(cr)}
\arguments{\item{\code{cr}}{[\code{\link{cairo_t}}]  a cairo context}}
\value{[\code{\link{cairo_line_join_t}}]  the current line join style.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
