\alias{cairoGetLineWidth}
\name{cairoGetLineWidth}
\title{cairoGetLineWidth}
\description{Gets the current line width, as set by \code{\link{cairoSetLineWidth}}.}
\usage{cairoGetLineWidth(cr)}
\arguments{\item{\code{cr}}{[\code{\link{Cairo}}]  a cairo context}}
\value{[numeric]  the current line width, in user-space units.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
