\alias{cairoGetMatrix}
\name{cairoGetMatrix}
\title{cairoGetMatrix}
\description{Stores the current transformation matrix (CTM) into \code{matrix}.}
\usage{cairoGetMatrix(cr, matrix)}
\arguments{
\item{\code{cr}}{[\code{\link{Cairo}}]  a cairo context}
\item{\code{matrix}}{[\code{\link{CairoMatrix}}]  return value for the matrix}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
