\alias{cairoGetMiterLimit}
\name{cairoGetMiterLimit}
\title{cairoGetMiterLimit}
\description{Gets the current miter limit, as set by \code{\link{cairoSetMiterLimit}}.}
\usage{cairoGetMiterLimit(cr)}
\arguments{\item{\verb{cr}}{[\code{\link{Cairo}}]  a cairo context}}
\value{[numeric]  the current miter limit.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
