\alias{cairoGetOperator}
\name{cairoGetOperator}
\title{cairoGetOperator}
\description{Gets the current compositing operator for a cairo context.}
\usage{cairoGetOperator(cr)}
\arguments{\item{\code{cr}}{[\code{\link{cairo_t}}]  a cairo context}}
\value{[\code{\link{cairo_operator_t}}]  the current compositing operator.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
