\alias{cairoGetSource}
\name{cairoGetSource}
\title{cairoGetSource}
\description{Gets the current source pattern for \code{cr}.}
\usage{cairoGetSource(cr)}
\arguments{\item{\code{cr}}{[\code{\link{cairo_t}}]  a cairo context}}
\value{[\code{\link{cairo_pattern_t}}]  the current source pattern.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
