\alias{cairoGlyphPath}
\name{cairoGlyphPath}
\title{cairoGlyphPath}
\description{Adds closed paths for the glyphs to the current path.  The generated
path if filled, achieves an effect similar to that of
\code{\link{cairoShowGlyphs}}.}
\usage{cairoGlyphPath(cr, glyphs)}
\arguments{
\item{\verb{cr}}{[\code{\link{Cairo}}]  a cairo context}
\item{\verb{glyphs}}{[\code{\link{CairoGlyph}}]  list of glyphs to show}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
