\alias{cairoImageSurfaceGetFormat}
\name{cairoImageSurfaceGetFormat}
\title{cairoImageSurfaceGetFormat}
\description{Get the format of the surface.}
\usage{cairoImageSurfaceGetFormat(surface)}
\arguments{\item{\verb{surface}}{[\code{\link{CairoSurface}}]  a \verb{cairo_image_surface_t}}}
\details{  Since 1.2}
\value{[\code{\link{CairoFormat}}]  the format of the surface}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
