\alias{cairoMatrixInitIdentity}
\name{cairoMatrixInitIdentity}
\title{cairoMatrixInitIdentity}
\description{Modifies \code{matrix} to be an identity transformation.}
\usage{cairoMatrixInitIdentity()}
\value{
A list containing the following elements:
\item{\verb{matrix}}{[\code{\link{CairoMatrix}}]  a \code{\link{CairoMatrix}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
