\alias{cairoMatrixInitScale}
\name{cairoMatrixInitScale}
\title{cairoMatrixInitScale}
\description{Initializes \code{matrix} to a transformation that scales by \code{sx} and \code{sy}
in the X and Y dimensions, respectively.}
\usage{cairoMatrixInitScale(sx, sy)}
\arguments{
\item{\verb{sx}}{[numeric]  scale factor in the X direction}
\item{\verb{sy}}{[numeric]  scale factor in the Y direction}
}
\value{
A list containing the following elements:
\item{\verb{matrix}}{[\code{\link{CairoMatrix}}]  a \code{\link{CairoMatrix}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
