\alias{cairoMatrixTransformPoint}
\name{cairoMatrixTransformPoint}
\title{cairoMatrixTransformPoint}
\description{Transforms the point (\code{x}, \code{y}) by \code{matrix}.}
\usage{cairoMatrixTransformPoint(matrix, x, y)}
\arguments{
\item{\verb{matrix}}{[\code{\link{CairoMatrix}}]  a \code{\link{CairoMatrix}}}
\item{\verb{x}}{[numeric]  X position. An in/out parameter}
\item{\verb{y}}{[numeric]  Y position. An in/out parameter}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
