\alias{cairoMoveTo}
\name{cairoMoveTo}
\title{cairoMoveTo}
\description{If the current subpath is not empty, begin a new subpath. After
this call the current point will be (\code{x}, \code{y}).}
\usage{cairoMoveTo(cr, x, y)}
\arguments{
\item{\code{cr}}{[\code{\link{cairo_t}}]  a cairo context}
\item{\code{x}}{[numeric]  the X coordinate of the new position}
\item{\code{y}}{[numeric]  the Y coordinate of the new position}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
