\alias{cairoMoveTo}
\name{cairoMoveTo}
\title{cairoMoveTo}
\description{Begin a new sub-path. After this call the current point will be (\code{x},
\code{y}).}
\usage{cairoMoveTo(cr, x, y)}
\arguments{
\item{\verb{cr}}{[\code{\link{Cairo}}]  a cairo context}
\item{\verb{x}}{[numeric]  the X coordinate of the new position}
\item{\verb{y}}{[numeric]  the Y coordinate of the new position}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
