\alias{cairoNewPath}
\name{cairoNewPath}
\title{cairoNewPath}
\description{Clears the current path. After this call there will be no current
point.}
\usage{cairoNewPath(cr)}
\arguments{\item{\code{cr}}{[\code{\link{cairo_t}}]  a cairo context}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
