\alias{cairoNewPath}
\name{cairoNewPath}
\title{cairoNewPath}
\description{Clears the current path. After this call there will be no path and
no current point.}
\usage{cairoNewPath(cr)}
\arguments{\item{\verb{cr}}{[\code{\link{Cairo}}]  a cairo context}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
