\alias{cairoPatternGetExtend}
\name{cairoPatternGetExtend}
\title{cairoPatternGetExtend}
\description{\emph{undocumented
}}
\usage{cairoPatternGetExtend(pattern)}
\arguments{\item{\code{pattern}}{[\code{\link{cairo_pattern_t}}] }}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
