\alias{cairoPatternGetExtend}
\name{cairoPatternGetExtend}
\title{cairoPatternGetExtend}
\description{Gets the current extend mode for a pattern.  See \code{\link{CairoExtend}}
for details on the semantics of each extend strategy.}
\usage{cairoPatternGetExtend(pattern)}
\arguments{\item{\verb{pattern}}{[\code{\link{CairoPattern}}]  a \code{\link{CairoPattern}}}}
\value{[\code{\link{CairoExtend}}]  the current extend strategy used for drawing the
pattern.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
