\alias{cairoPatternGetMatrix}
\name{cairoPatternGetMatrix}
\title{cairoPatternGetMatrix}
\description{Stores the pattern's transformation matrix into \code{matrix}.}
\usage{cairoPatternGetMatrix(pattern, matrix)}
\arguments{
\item{\verb{pattern}}{[\verb{\link{CairoPattern}}]  a \verb{\link{CairoPattern}}}
\item{\verb{matrix}}{[\verb{\link{CairoMatrix}}]  return value for the matrix}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
