\alias{cairoPatternGetType}
\name{cairoPatternGetType}
\title{cairoPatternGetType}
\description{This function returns the type a pattern.
See \code{\link{CairoPatternType}} for available types.}
\usage{cairoPatternGetType(pattern)}
\arguments{\item{\verb{pattern}}{[\code{\link{CairoPattern}}]  a \code{\link{CairoPattern}}}}
\details{  Since 1.2}
\value{[\code{\link{CairoPatternType}}]  The type of \code{pattern}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
