\alias{cairoPatternSetExtend}
\name{cairoPatternSetExtend}
\title{cairoPatternSetExtend}
\usage{cairoPatternSetExtend(pattern, extend)}
\arguments{
\item{\code{pattern}}{[\code{\link{cairo_pattern_t}}] }
\item{\code{extend}}{[\code{\link{cairo_extend_t}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
