\alias{cairoPatternSetExtend}
\name{cairoPatternSetExtend}
\title{cairoPatternSetExtend}
\description{\emph{undocumented
}}
\usage{cairoPatternSetExtend(pattern, extend)}
\arguments{
\item{\code{pattern}}{[\code{\link{CairoPattern}}] }
\item{\code{extend}}{[\code{\link{CairoExtend}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
