\alias{cairoPatternSetFilter}
\name{cairoPatternSetFilter}
\title{cairoPatternSetFilter}
\description{Sets the filter to be used for resizing when using this pattern.
See \code{\link{CairoFilter}} for details on each filter.}
\usage{cairoPatternSetFilter(pattern, filter)}
\arguments{
\item{\code{pattern}}{[\code{\link{CairoPattern}}]  a \code{\link{CairoPattern}}}
\item{\code{filter}}{[\code{\link{CairoFilter}}]  a \code{\link{CairoFilter}} describing the filter to use for resizing
the pattern}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
