\alias{cairoPatternSetFilter}
\name{cairoPatternSetFilter}
\title{cairoPatternSetFilter}
\description{\emph{undocumented
}}
\usage{cairoPatternSetFilter(pattern, filter)}
\arguments{
\item{\code{pattern}}{[\code{\link{CairoPattern}}] }
\item{\code{filter}}{[\code{\link{CairoFilter}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
