\alias{cairoPatternStatus}
\name{cairoPatternStatus}
\title{cairoPatternStatus}
\description{Checks whether an error has previously occurred for this
pattern.}
\usage{cairoPatternStatus(pattern)}
\arguments{\item{\verb{pattern}}{[\code{\link{CairoPattern}}]  a \code{\link{CairoPattern}}}}
\value{[\code{\link{CairoStatus}}] \code{CAIRO_STATUS_SUCCESS}, \code{CAIRO_STATUS_NO_MEMORY}, or
\code{CAIRO_STATUS_PATTERN_TYPE_MISMATCH}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
