\alias{cairoPsGetLevels}
\name{cairoPsGetLevels}
\title{cairoPsGetLevels}
\description{Used to retrieve the list of supported levels. See
\code{\link{cairoPsSurfaceRestrictToLevel}}.}
\usage{cairoPsGetLevels()}
\details{  Since 1.6}
\value{
A list containing the following elements:
\item{\verb{levels}}{[\code{\link{CairoPsLevel}}]  supported level list}
\item{\verb{nlevels}}{[integer]  list length}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
