\alias{cairoRestore}
\name{cairoRestore}
\title{cairoRestore}
\description{Restores \code{cr} to the state saved by a preceding call to
\code{\link{cairoSave}} and removes that state from the stack of
saved states.}
\usage{cairoRestore(cr)}
\arguments{\item{\verb{cr}}{[\code{\link{Cairo}}]  a \code{\link{Cairo}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
