\alias{cairoScaledFontExtents}
\name{cairoScaledFontExtents}
\title{cairoScaledFontExtents}
\description{Gets the metrics for a \verb{\link{CairoScaledFont}}.}
\usage{cairoScaledFontExtents(scaled.font)}
\arguments{\item{\verb{scaled.font}}{[\verb{\link{CairoScaledFont}}]  a \verb{\link{CairoScaledFont}}}}
\value{
A list containing the following elements:
\item{\verb{extents}}{[\verb{\link{CairoFontExtents}}]  a \verb{\link{CairoFontExtents}} which to store the retrieved extents.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
