\alias{cairoScaledFontStatus}
\name{cairoScaledFontStatus}
\title{cairoScaledFontStatus}
\description{Checks whether an error has previously occurred for this
scaled_font.}
\usage{cairoScaledFontStatus(scaled.font)}
\arguments{\item{\code{scaled.font}}{[\code{\link{cairo_scaled_font_t}}]  a \code{\link{cairo_scaled_font_t}}}}
\value{[\code{\link{cairo_status_t}}]  \code{CAIRO_STATUS_SUCCESS} or another error such as
  \code{CAIRO_STATUS_NO_MEMORY}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
