\alias{cairoScaledFontStatus}
\name{cairoScaledFontStatus}
\title{cairoScaledFontStatus}
\description{Checks whether an error has previously occurred for this
scaled_font.}
\usage{cairoScaledFontStatus(scaled.font)}
\arguments{\item{\verb{scaled.font}}{[\verb{\link{CairoScaledFont}}]  a \verb{\link{CairoScaledFont}}}}
\value{[\verb{\link{CairoStatus}}] \code{CAIRO_STATUS_SUCCESS} or another error such as
  \code{CAIRO_STATUS_NO_MEMORY}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
