\alias{cairoSetMatrix}
\name{cairoSetMatrix}
\title{cairoSetMatrix}
\description{Modifies the current transformation matrix (CTM) by setting it
equal to \code{matrix}.}
\usage{cairoSetMatrix(cr, matrix)}
\arguments{
\item{\verb{cr}}{[\verb{\link{Cairo}}]  a cairo context}
\item{\verb{matrix}}{[\verb{\link{CairoMatrix}}]  a transformation matrix from user space to device space}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
