\alias{cairoShowGlyphs}
\name{cairoShowGlyphs}
\title{cairoShowGlyphs}
\description{\emph{undocumented
}}
\usage{cairoShowGlyphs(cr, glyphs, num.glyphs)}
\arguments{
\item{\code{cr}}{[\code{\link{Cairo}}] }
\item{\code{glyphs}}{[\code{\link{CairoGlyph}}] }
\item{\code{num.glyphs}}{[integer] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
