\alias{cairoShowGlyphs}
\name{cairoShowGlyphs}
\title{cairoShowGlyphs}
\usage{cairoShowGlyphs(cr, glyphs, num.glyphs)}
\arguments{
\item{\code{cr}}{[\code{\link{cairo_t}}] }
\item{\code{glyphs}}{[\code{\link{cairo_glyph_t}}] }
\item{\code{num.glyphs}}{[integer] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
