\alias{cairoShowGlyphs}
\name{cairoShowGlyphs}
\title{cairoShowGlyphs}
\description{A drawing operator that generates the shape from a list of glyphs,
rendered according to the current font face, font size
(font matrix), and font options.}
\usage{cairoShowGlyphs(cr, glyphs, num.glyphs)}
\arguments{
\item{\verb{cr}}{[\code{\link{Cairo}}]  a cairo context}
\item{\verb{glyphs}}{[\code{\link{CairoGlyph}}]  list of glyphs to show}
\item{\verb{num.glyphs}}{[integer]  number of glyphs to show}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
