\alias{cairoStatus}
\name{cairoStatus}
\title{cairoStatus}
\description{Checks whether an error has previously occurred for this context.}
\usage{cairoStatus(cr)}
\arguments{\item{\code{cr}}{[\code{\link{cairo_t}}]  a cairo context}}
\value{[\code{\link{cairo_status_t}}] the current status of this context, see \code{\link{cairo_status_t}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
