\alias{cairoStatus}
\name{cairoStatus}
\title{cairoStatus}
\description{Checks whether an error has previously occurred for this context.}
\usage{cairoStatus(cr)}
\arguments{\item{\verb{cr}}{[\code{\link{Cairo}}]  a cairo context}}
\value{[\code{\link{CairoStatus}}]  the current status of this context, see \code{\link{CairoStatus}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
