\alias{cairoStatusToString}
\name{cairoStatusToString}
\title{cairoStatusToString}
\description{Provides a human-readable description of a \code{\link{cairo_status_t}}.}
\usage{cairoStatusToString(status)}
\arguments{\item{\code{status}}{[\code{\link{cairo_status_t}}]  a cairo status}}
\value{[char] a string representation of the status}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
