\alias{cairoSurfaceDestroy}
\name{cairoSurfaceDestroy}
\title{cairoSurfaceDestroy}
\description{Decreases the reference count on \code{surface} by one.  See
\code{\link{cairoSurfaceReference}}.}
\usage{cairoSurfaceDestroy(surface)}
\arguments{\item{\code{surface}}{[\code{\link{CairoSurface}}]  a \code{\link{Cairo}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
