\alias{cairoSurfaceDestroy}
\name{cairoSurfaceDestroy}
\title{cairoSurfaceDestroy}
\description{Decreases the reference count on \code{surface} by one. If the result is
zero, then \code{surface} and all associated resources are freed.}
\usage{cairoSurfaceDestroy(surface)}
\arguments{\item{\verb{surface}}{[\code{\link{CairoSurface}}]  a \code{\link{CairoSurface}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
