\alias{cairoSurfaceGetContent}
\name{cairoSurfaceGetContent}
\title{cairoSurfaceGetContent}
\description{This function returns the content type of \code{surface} which indicates
whether the surface contains color and/or alpha information. See
\verb{\link{CairoContent}}.}
\usage{cairoSurfaceGetContent(surface)}
\arguments{\item{\verb{surface}}{[\verb{\link{CairoSurface}}]  a \verb{\link{CairoSurface}}}}
\details{  Since  1.2}
\value{[\verb{\link{CairoContent}}]  The content type of \code{surface}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
