\alias{cairoSurfaceShowPage}
\name{cairoSurfaceShowPage}
\title{cairoSurfaceShowPage}
\description{Emits and clears the current page for backends that support multiple
pages.  Use \code{\link{cairoSurfaceCopyPage}} if you don't want to clear the page.}
\usage{cairoSurfaceShowPage(surface)}
\arguments{\item{\code{surface}}{[\code{\link{CairoSurface}}]  a \code{cairo_Surface_t}}}
\details{  Since  1.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
