\alias{cairoSvgGetVersions}
\name{cairoSvgGetVersions}
\title{cairoSvgGetVersions}
\description{Used to retrieve the list of supported versions. See
\code{\link{cairoSvgSurfaceRestrictToVersion}}.}
\usage{cairoSvgGetVersions(versions, num.versions)}
\arguments{
\item{\verb{versions}}{[\verb{\link{CairoSvgVersion}}]  supported version list}
\item{\verb{num.versions}}{[integer]  list length}
}
\details{  Since  1.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
