\alias{cairoSvgGetVersions}
\name{cairoSvgGetVersions}
\title{cairoSvgGetVersions}
\description{Used to retrieve the list of supported versions. See
\code{\link{cairoSvgSurfaceRestrictToVersion}}.}
\usage{cairoSvgGetVersions()}
\details{  Since 1.2}
\value{
A list containing the following elements:
\item{\verb{versions}}{[\code{\link{CairoSvgVersion}}]  supported version list}
\item{\verb{num.versions}}{[integer]  list length}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
