\alias{gdkCairoRectangle}
\name{gdkCairoRectangle}
\title{gdkCairoRectangle}
\description{Adds the given rectangle to the current path of \code{cr}.}
\usage{gdkCairoRectangle(cr, rectangle)}
\arguments{
\item{\verb{cr}}{a \code{\link{Cairo}}}
\item{\verb{rectangle}}{a \code{\link{GdkRectangle}}}
}
\details{Since 2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
