\alias{gdkCairoRegion}
\name{gdkCairoRegion}
\title{gdkCairoRegion}
\description{Adds the given region to the current path of \code{cr}.}
\usage{gdkCairoRegion(cr, region)}
\arguments{
\item{\verb{cr}}{[\verb{\link{Cairo}}]  a \verb{\link{Cairo}}}
\item{\verb{region}}{[\verb{\link{GdkRegion}}]  a \verb{\link{GdkRegion}}}
}
\details{  Since  2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
