\alias{gdkCairoRegion}
\name{gdkCairoRegion}
\title{gdkCairoRegion}
\description{Adds the given region to the current path of \code{cr}.}
\usage{gdkCairoRegion(cr, region)}
\arguments{
\item{\verb{cr}}{a \code{\link{Cairo}}}
\item{\verb{region}}{a \code{\link{GdkRegion}}}
}
\details{Since 2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
