\alias{gdkCairoSetSourceColor}
\name{gdkCairoSetSourceColor}
\title{gdkCairoSetSourceColor}
\description{Sets the specified \code{\link{GdkColor}} as the source color of \code{cr}.}
\usage{gdkCairoSetSourceColor(cr, color)}
\arguments{
\item{\code{cr}}{[\code{\link{Cairo}}]  a \code{\link{Cairo}}}
\item{\code{color}}{[\code{\link{GdkColor}}]  a \code{\link{GdkColor}}}
}
\details{  Since  2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
