\alias{gdkColormapFreeColors}
\name{gdkColormapFreeColors}
\title{gdkColormapFreeColors}
\description{Frees previously allocated colors.}
\usage{gdkColormapFreeColors(object, colors)}
\arguments{
\item{\code{object}}{[\code{\link{GdkColormap}}]  a \code{\link{GdkColormap}}.}
\item{\code{colors}}{[\code{\link{GdkColor}}]  the colors to free.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
