\alias{gdkColormapFreeColors}
\name{gdkColormapFreeColors}
\title{gdkColormapFreeColors}
\description{Frees previously allocated colors.}
\usage{gdkColormapFreeColors(object, colors)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkColormap}}.}
\item{\verb{colors}}{the colors to free.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
