\alias{gdkColormapGetScreen}
\name{gdkColormapGetScreen}
\title{gdkColormapGetScreen}
\description{Gets the screen for which this colormap was created.}
\usage{gdkColormapGetScreen(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkColormap}}]  a \code{\link{GdkColormap}}}}
\details{  Since  2.2}
\value{[\code{\link{GdkScreen}}]  the screen for which this colormap was created.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
