\alias{gdkColormapGetSystem}
\name{gdkColormapGetSystem}
\title{gdkColormapGetSystem}
\description{Gets the system's default colormap for the default screen. (See
\code{gdkColormapGetSystemForScreen()})}
\usage{gdkColormapGetSystem()}
\value{[\verb{\link{GdkColormap}}]  the default colormap.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
