\alias{gdkColormapGetVisual}
\name{gdkColormapGetVisual}
\title{gdkColormapGetVisual}
\description{Returns the visual for which a given colormap was created.}
\usage{gdkColormapGetVisual(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkColormap}}.}}
\value{[\code{\link{GdkVisual}}] the visual of the colormap.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
