\alias{gdkCursorGetDisplay}
\name{gdkCursorGetDisplay}
\title{gdkCursorGetDisplay}
\description{Returns the display on which the \code{\link{GdkCursor}} is defined.}
\usage{gdkCursorGetDisplay(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkCursor}}.}}
\details{Since 2.2}
\value{[\code{\link{GdkDisplay}}] the \code{\link{GdkDisplay}} associated to \code{cursor}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
