\alias{gdkDeviceGetCorePointer}
\name{gdkDeviceGetCorePointer}
\title{gdkDeviceGetCorePointer}
\description{Returns the core pointer device for the default display.}
\usage{gdkDeviceGetCorePointer()}
\value{[\verb{\link{GdkDevice}}] .}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
