\alias{gdkDeviceSetAxisUse}
\name{gdkDeviceSetAxisUse}
\title{gdkDeviceSetAxisUse}
\description{Specifies how an axis of a device is used.}
\usage{gdkDeviceSetAxisUse(object, index, use)}
\arguments{
\item{\code{object}}{[\code{\link{GdkDevice}}] a \code{\link{GdkDevice}}.}
\item{\code{index}}{[numeric] the index of the axis.}
\item{\code{use}}{[\code{\link{GdkAxisUse}}] specifies how the axis is used.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
