\alias{gdkDeviceSetKey}
\name{gdkDeviceSetKey}
\title{gdkDeviceSetKey}
\description{Specifies the X key event to generate when a function button of a device
is pressed.}
\usage{gdkDeviceSetKey(object, index, keyval, modifiers)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkDevice}}.}
\item{\verb{index}}{the index of the function button to set.}
\item{\verb{keyval}}{the keyval to generate.}
\item{\verb{modifiers}}{the modifiers to set.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
