\alias{gdkDeviceSetSource}
\name{gdkDeviceSetSource}
\title{gdkDeviceSetSource}
\description{Sets the source type for an input device.}
\usage{gdkDeviceSetSource(object, source)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkDevice}}] a \verb{\link{GdkDevice}}.}
\item{\verb{source}}{[\code{\link{GdkInputSource}}] the source type.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
