\alias{gdkDeviceSetSource}
\name{gdkDeviceSetSource}
\title{gdkDeviceSetSource}
\description{Sets the source type for an input device.}
\usage{gdkDeviceSetSource(object, source)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkDevice}}.}
\item{\verb{source}}{the source type.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
