\alias{gdkDevicesList}
\name{gdkDevicesList}
\title{gdkDevicesList}
\description{Returns the list of available input devices for the default display.
The list is statically allocated and should not be freed.}
\usage{gdkDevicesList()}
\value{[list] a list of \code{\link{GdkDevice}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
