\alias{gdkDisplayClose}
\name{gdkDisplayClose}
\title{gdkDisplayClose}
\description{Closes the connection to the windowing system for the given display,
and cleans up associated resources.}
\usage{gdkDisplayClose(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkDisplay}}}}
\details{Since 2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
