\alias{gdkDisplayGetCorePointer}
\name{gdkDisplayGetCorePointer}
\title{gdkDisplayGetCorePointer}
\description{Returns the core pointer device for the given display}
\usage{gdkDisplayGetCorePointer(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkDisplay}}}}
\details{Since 2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
