\alias{gdkDisplayGetDefaultScreen}
\name{gdkDisplayGetDefaultScreen}
\title{gdkDisplayGetDefaultScreen}
\description{Get the default \code{\link{GdkScreen}} for \code{display}.}
\usage{gdkDisplayGetDefaultScreen(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkDisplay}}}}
\details{Since 2.2}
\value{[\code{\link{GdkScreen}}] the default \code{\link{GdkScreen}} object for \code{display}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
