\alias{gdkDisplayGetMaximalCursorSize}
\name{gdkDisplayGetMaximalCursorSize}
\title{gdkDisplayGetMaximalCursorSize}
\description{Gets the maximal size to use for cursors on \code{display}.}
\usage{gdkDisplayGetMaximalCursorSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkDisplay}}}}
\details{Since 2.4}
\value{
A list containing the following elements:
\item{\verb{width}}{the return location for the maximal cursor width}
\item{\verb{height}}{the return location for the maximal cursor height}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
