\alias{gdkDisplayGetName}
\name{gdkDisplayGetName}
\title{gdkDisplayGetName}
\description{Gets the name of the display.}
\usage{gdkDisplayGetName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkDisplay}}}}
\details{Since 2.2}
\value{[character] a string representing the display name. This string is owned
by GDK and should not be modified or freed.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
