\alias{gdkDisplayGetScreen}
\name{gdkDisplayGetScreen}
\title{gdkDisplayGetScreen}
\description{Returns a screen object for one of the screens of the display.}
\usage{gdkDisplayGetScreen(object, screen.num)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkDisplay}}]  a \verb{\link{GdkDisplay}}}
\item{\verb{screen.num}}{[integer]  the screen number}
}
\details{  Since  2.2}
\value{[\verb{\link{GdkScreen}}]  the \verb{\link{GdkScreen}} object}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
