\alias{gdkDisplayGetScreen}
\name{gdkDisplayGetScreen}
\title{gdkDisplayGetScreen}
\description{Returns a screen object for one of the screens of the display.}
\usage{gdkDisplayGetScreen(object, screen.num)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkDisplay}}}
\item{\verb{screen.num}}{the screen number}
}
\details{Since 2.2}
\value{[\code{\link{GdkScreen}}] the \code{\link{GdkScreen}} object}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
