\alias{gdkDisplayKeyboardUngrab}
\name{gdkDisplayKeyboardUngrab}
\title{gdkDisplayKeyboardUngrab}
\description{Release any keyboard grab}
\usage{gdkDisplayKeyboardUngrab(object, time. = "GDK_CURRENT_TIME")}
\arguments{
\item{\verb{object}}{a \code{\link{GdkDisplay}}.}
\item{\verb{time.}}{a timestap (e.g \verb{GDK_CURRENT_TIME}).}
}
\details{Since 2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
