\alias{gdkDisplayListDevices}
\name{gdkDisplayListDevices}
\title{gdkDisplayListDevices}
\description{Returns the list of available input devices attached to \code{display}.}
\usage{gdkDisplayListDevices(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkDisplay}}]  a \verb{\link{GdkDisplay}}}}
\details{  Since  2.2}
\value{[list]  a list of \verb{\link{GdkDevice}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
