\alias{gdkDisplayListDevices}
\name{gdkDisplayListDevices}
\title{gdkDisplayListDevices}
\description{Returns the list of available input devices attached to \code{display}.
The list is statically allocated and should not be freed.}
\usage{gdkDisplayListDevices(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkDisplay}}}}
\details{Since 2.2}
\value{[list] a list of \code{\link{GdkDevice}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
