\alias{gdkDisplayManagerGet}
\name{gdkDisplayManagerGet}
\title{gdkDisplayManagerGet}
\description{Returns the global \verb{\link{GdkDisplayManager}} singleton; \code{gdkParsePargs()},
\code{gdkInit()}, or \code{gdkInitCheck()} must have been called first.}
\usage{gdkDisplayManagerGet()}
\details{  Since  2.2}
\value{[\verb{\link{GdkDisplayManager}}]  the singleton \verb{\link{GdkDisplayManager}} object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
