\alias{gdkDisplayManagerGet}
\name{gdkDisplayManagerGet}
\title{gdkDisplayManagerGet}
\description{Returns the global \code{\link{GdkDisplayManager}} singleton; \code{gdk\Parse\Pargs()},
\code{gdk\Init()}, or \code{gdk\Init\Check()} must have been called first.}
\usage{gdkDisplayManagerGet()}
\details{  Since  2.2}
\value{[\code{\link{GdkDisplayManager}}]  the singleton \code{\link{GdkDisplayManager}} object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
