\alias{gdkDisplayManagerGet}
\name{gdkDisplayManagerGet}
\title{gdkDisplayManagerGet}
\description{Gets the singleton \code{\link{GdkDisplayManager}} object.}
\usage{gdkDisplayManagerGet()}
\details{Since 2.2}
\value{[\code{\link{GdkDisplayManager}}] The global \code{\link{GdkDisplayManager}} singleton; \code{gdkParsePargs()},
\code{gdkInit()}, or \code{gdkInitCheck()} must have been called first. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
