\alias{gdkDisplayManagerGetDefaultDisplay}
\name{gdkDisplayManagerGetDefaultDisplay}
\title{gdkDisplayManagerGetDefaultDisplay}
\description{Gets the default \verb{\link{GdkDisplay}}.}
\usage{gdkDisplayManagerGetDefaultDisplay(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkDisplayManager}}]  a \verb{\link{GdkDisplayManager}} }}
\details{  Since  2.2}
\value{[\verb{\link{GdkDisplay}}]  a \verb{\link{GdkDisplay}}, or \code{NULL} if there is no default
  display.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
