\alias{gdkDisplayManagerGetDefaultDisplay}
\name{gdkDisplayManagerGetDefaultDisplay}
\title{gdkDisplayManagerGetDefaultDisplay}
\description{Gets the default \code{\link{GdkDisplay}}.}
\usage{gdkDisplayManagerGetDefaultDisplay(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkDisplayManager}}}}
\details{Since 2.2}
\value{[\code{\link{GdkDisplay}}] a \code{\link{GdkDisplay}}, or \code{NULL} if there is no default
display. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
