\alias{gdkDisplayManagerListDisplays}
\name{gdkDisplayManagerListDisplays}
\title{gdkDisplayManagerListDisplays}
\description{List all currently open displays.}
\usage{gdkDisplayManagerListDisplays(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkDisplayManager}}}}
\details{Since 2.2}
\value{[list] a newly allocated \verb{list} of \code{\link{GdkDisplay}} objects.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
