\alias{gdkDisplayPointerUngrab}
\name{gdkDisplayPointerUngrab}
\title{gdkDisplayPointerUngrab}
\description{Release any pointer grab.}
\usage{gdkDisplayPointerUngrab(object, time. = "GDK_CURRENT_TIME")}
\arguments{
\item{\verb{object}}{a \code{\link{GdkDisplay}}.}
\item{\verb{time.}}{a timestap (e.g. \code{GDK_CURRENT_TIME}).}
}
\details{Since 2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
