\alias{gdkDisplayPutEvent}
\name{gdkDisplayPutEvent}
\title{gdkDisplayPutEvent}
\description{Appends a copy of the given event onto the front of the event
queue for \code{display}.}
\usage{gdkDisplayPutEvent(object, event)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkDisplay}}]  a \verb{\link{GdkDisplay}}}
\item{\verb{event}}{[\verb{\link{GdkEvent}}]  a \verb{\link{GdkEvent}}.}
}
\details{  Since  2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
