\alias{gdkDisplaySupportsSelectionNotification}
\name{gdkDisplaySupportsSelectionNotification}
\title{gdkDisplaySupportsSelectionNotification}
\description{Returns whether \code{\link{GdkEventOwnerChange}} events will be 
sent when the owner of a selection changes.}
\usage{gdkDisplaySupportsSelectionNotification(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkDisplay}}}}
\details{Since 2.6}
\value{[logical] whether \code{\link{GdkEventOwnerChange}} events will
be sent.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
