\alias{gdkDragAbort}
\name{gdkDragAbort}
\title{gdkDragAbort}
\description{Aborts a drag without dropping. }
\usage{gdkDragAbort(object, time)}
\arguments{
\item{\code{object}}{[\code{\link{GdkDragContext}}]  a \code{\link{GdkDragContext}}.}
\item{\code{time}}{[numeric]  the timestamp for this operation.}
}
\details{This function is called by the drag source.  }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
