\alias{gdkDragContextNew}
\name{gdkDragContextNew}
\title{gdkDragContextNew}
\description{Creates a new \code{\link{GdkDragContext}}.}
\usage{gdkDragContextNew()}
\value{[\code{\link{GdkDragContext}}] the newly created \code{\link{GdkDragContext}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
