\alias{gdkDragDrop}
\name{gdkDragDrop}
\title{gdkDragDrop}
\description{Drops on the current destination.}
\usage{gdkDragDrop(object, time)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkDragContext}}.}
\item{\verb{time}}{the timestamp for this operation.}
}
\details{This function is called by the drag source.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
