\alias{gdkDrawableGetColormap}
\name{gdkDrawableGetColormap}
\title{gdkDrawableGetColormap}
\description{Gets the colormap for \code{drawable}, if one is set; returns
\code{NULL} otherwise.}
\usage{gdkDrawableGetColormap(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkDrawable}}}}
\value{[\code{\link{GdkColormap}}] the colormap, or \code{NULL}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
