\alias{gdkDrawableGetDisplay}
\name{gdkDrawableGetDisplay}
\title{gdkDrawableGetDisplay}
\description{Gets the \code{\link{GdkDisplay}} associated with a \code{\link{GdkDrawable}}.}
\usage{gdkDrawableGetDisplay(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkDrawable}}}}
\details{Since 2.2}
\value{[\code{\link{GdkDisplay}}] the \code{\link{GdkDisplay}} associated with \code{drawable}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
