\alias{gdkDrawableGetScreen}
\name{gdkDrawableGetScreen}
\title{gdkDrawableGetScreen}
\description{Gets the \code{\link{GdkScreen}} associated with a \code{\link{GdkDrawable}}.}
\usage{gdkDrawableGetScreen(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkDrawable}}}}
\details{Since 2.2}
\value{[\code{\link{GdkScreen}}] the \code{\link{GdkScreen}} associated with \code{drawable}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
