\alias{gdkDrawableGetVisual}
\name{gdkDrawableGetVisual}
\title{gdkDrawableGetVisual}
\description{Gets the \verb{\link{GdkVisual}} describing the pixel format of \code{drawable}.}
\usage{gdkDrawableGetVisual(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkDrawable}}]  a \verb{\link{GdkDrawable}}}}
\value{[\verb{\link{GdkVisual}}]  a \verb{\link{GdkVisual}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
