\alias{gdkDrawableGetVisual}
\name{gdkDrawableGetVisual}
\title{gdkDrawableGetVisual}
\description{Gets the \code{\link{GdkVisual}} describing the pixel format of \code{drawable}.}
\usage{gdkDrawableGetVisual(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkDrawable}}}}
\value{[\code{\link{GdkVisual}}] a \code{\link{GdkVisual}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
