\alias{gdkDrawLine}
\name{gdkDrawLine}
\title{gdkDrawLine}
\description{Draws a line, using the foreground color and other attributes of 
the \verb{\link{GdkGC}}.}
\usage{gdkDrawLine(object, gc, x1, y1, x2, y2)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkDrawable}}]  a \verb{\link{GdkDrawable}} (a \verb{\link{GdkWindow}} or a \verb{\link{GdkPixmap}}). }
\item{\verb{gc}}{[\verb{\link{GdkGC}}]  a \verb{\link{GdkGC}}.}
\item{\verb{x1}}{[integer]  the x coordinate of the start point.}
\item{\verb{y1}}{[integer]  the y coordinate of the start point.}
\item{\verb{x2}}{[integer]  the x coordinate of the end point.}
\item{\verb{y2}}{[integer]  the y coordinate of the end point.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
