\alias{gdkDrawLine}
\name{gdkDrawLine}
\title{gdkDrawLine}
\description{Draws a line, using the foreground color and other attributes of 
the \code{\link{GdkGC}}.}
\usage{gdkDrawLine(object, gc, x1, y1, x2, y2)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkDrawable}} (a \code{\link{GdkWindow}} or a \code{\link{GdkPixmap}}).}
\item{\verb{gc}}{a \code{\link{GdkGC}}.}
\item{\verb{x1}}{the x coordinate of the start point.}
\item{\verb{y1}}{the y coordinate of the start point.}
\item{\verb{x2}}{the x coordinate of the end point.}
\item{\verb{y2}}{the y coordinate of the end point.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
