\alias{gdkDrawPoint}
\name{gdkDrawPoint}
\title{gdkDrawPoint}
\description{Draws a point, using the foreground color and other attributes of 
the \verb{\link{GdkGC}}.}
\usage{gdkDrawPoint(object, gc, x, y)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkDrawable}}]  a \verb{\link{GdkDrawable}} (a \verb{\link{GdkWindow}} or a \verb{\link{GdkPixmap}}).}
\item{\verb{gc}}{[\verb{\link{GdkGC}}]  a \verb{\link{GdkGC}}.}
\item{\verb{x}}{[integer]  the x coordinate of the point.}
\item{\verb{y}}{[integer]  the y coordinate of the point.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
