\alias{gdkDrawPoint}
\name{gdkDrawPoint}
\title{gdkDrawPoint}
\description{Draws a point, using the foreground color and other attributes of 
the \code{\link{GdkGC}}.}
\usage{gdkDrawPoint(object, gc, x, y)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkDrawable}} (a \code{\link{GdkWindow}} or a \code{\link{GdkPixmap}}).}
\item{\verb{gc}}{a \code{\link{GdkGC}}.}
\item{\verb{x}}{the x coordinate of the point.}
\item{\verb{y}}{the y coordinate of the point.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
