\alias{gdkDrawPoints}
\name{gdkDrawPoints}
\title{gdkDrawPoints}
\description{Draws a number of points, using the foreground color and other 
attributes of the \code{\link{GdkGC}}.}
\usage{gdkDrawPoints(object, gc, points)}
\arguments{
\item{\code{object}}{[\code{\link{GdkDrawable}}]  a \code{\link{GdkDrawable}} (a \code{\link{GdkWindow}} or a \code{\link{GdkPixmap}}).}
\item{\code{gc}}{[\code{\link{GdkGC}}]  a \code{\link{GdkGC}}.}
\item{\code{points}}{[\code{\link{GdkPoint}}]  a list of \code{\link{GdkPoint}} structures.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
