\alias{gdkDrawPoints}
\name{gdkDrawPoints}
\title{gdkDrawPoints}
\description{Draws a number of points, using the foreground color and other 
attributes of the \verb{\link{GdkGC}}.}
\usage{gdkDrawPoints(object, gc, points)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkDrawable}}]  a \verb{\link{GdkDrawable}} (a \verb{\link{GdkWindow}} or a \verb{\link{GdkPixmap}}).}
\item{\verb{gc}}{[\verb{\link{GdkGC}}]  a \verb{\link{GdkGC}}.}
\item{\verb{points}}{[\verb{\link{GdkPoint}}]  a list of \verb{\link{GdkPoint}} structures.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
