\alias{gdkDrawPoints}
\name{gdkDrawPoints}
\title{gdkDrawPoints}
\description{Draws a number of points, using the foreground color and other 
attributes of the \code{\link{GdkGC}}.}
\usage{gdkDrawPoints(object, gc, points)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkDrawable}} (a \code{\link{GdkWindow}} or a \code{\link{GdkPixmap}}).}
\item{\verb{gc}}{a \code{\link{GdkGC}}.}
\item{\verb{points}}{a list of \code{\link{GdkPoint}} structures.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
