\alias{gdkDrawSegments}
\name{gdkDrawSegments}
\title{gdkDrawSegments}
\description{Draws a number of unconnected lines.}
\usage{gdkDrawSegments(object, gc, segs)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkDrawable}}]  a \verb{\link{GdkDrawable}} (a \verb{\link{GdkWindow}} or a \verb{\link{GdkPixmap}}).}
\item{\verb{gc}}{[\verb{\link{GdkGC}}]  a \verb{\link{GdkGC}}.}
\item{\verb{segs}}{[\verb{\link{GdkSegment}}]  a list of \verb{\link{GdkSegment}} structures specifying the start and 
  end points of the lines to be drawn.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
