\alias{gdkEventGetTime}
\name{gdkEventGetTime}
\title{gdkEventGetTime}
\description{returns \verb{GDK_CURRENT_TIME}. If \code{event} is \code{NULL}, returns \verb{GDK_CURRENT_TIME}.}
\usage{gdkEventGetTime(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkEvent}}}}
\value{[numeric] time stamp field from \code{event}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
