\alias{gdkEventNew}
\name{gdkEventNew}
\title{gdkEventNew}
\description{Creates a new event of the given type. All fields are set to 0.}
\usage{gdkEventNew(type)}
\arguments{\item{\verb{type}}{a \code{\link{GdkEventType}}}}
\details{Since 2.2}
\value{[\code{\link{GdkEvent}}] a newly-allocated \code{\link{GdkEvent}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
