\alias{gdkFlush}
\name{gdkFlush}
\title{gdkFlush}
\description{Flushes the X output buffer and waits until all requests have been processed
by the server. This is rarely needed by applications. It's main use is for
trapping X errors with \code{gdkErrorTrapPush()} and \code{gdkErrorTrapPop()}.}
\usage{gdkFlush()}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
