\alias{gdkFontGetDisplay}
\name{gdkFontGetDisplay}
\title{gdkFontGetDisplay}
\description{
Returns the \code{\link{GdkDisplay}} for \code{font}.
\strong{WARNING: \code{gdk_font_get_display} is deprecated and should not be used in newly-written code.}
}
\usage{gdkFontGetDisplay(object)}
\arguments{\item{\verb{object}}{the \code{\link{GdkFont}}.}}
\details{Since 2.2}
\value{[\code{\link{GdkDisplay}}] the corresponding \code{\link{GdkDisplay}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
