\alias{gdkGCGetScreen}
\name{gdkGCGetScreen}
\title{gdkGCGetScreen}
\description{Gets the \code{\link{GdkScreen}} for which \code{gc} was created}
\usage{gdkGCGetScreen(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkGC}}.}}
\details{Since 2.2}
\value{[\code{\link{GdkScreen}}] the \code{\link{GdkScreen}} for \code{gc}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
