\alias{gdkGCNew}
\name{gdkGCNew}
\title{gdkGCNew}
\description{Create a new graphics context with default values.}
\usage{gdkGCNew(drawable)}
\arguments{\item{\verb{drawable}}{a \code{\link{GdkDrawable}}. The created GC must always be used
with drawables of the same depth as this one.}}
\value{[\code{\link{GdkGC}}] the new graphics context.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
