\alias{gdkGCSetClipOrigin}
\name{gdkGCSetClipOrigin}
\title{gdkGCSetClipOrigin}
\description{Sets the origin of the clip mask. The coordinates are
interpreted relative to the upper-left corner of
the destination drawable of the current operation.}
\usage{gdkGCSetClipOrigin(object, x, y)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkGC}}]  a \verb{\link{GdkGC}}.}
\item{\verb{x}}{[integer]  the x-coordinate of the origin.}
\item{\verb{y}}{[integer]  the y-coordinate of the origin.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
