\alias{gdkGCSetFill}
\name{gdkGCSetFill}
\title{gdkGCSetFill}
\description{Set the fill mode for a graphics context.}
\usage{gdkGCSetFill(object, fill)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkGC}}]  a \verb{\link{GdkGC}}.}
\item{\verb{fill}}{[\code{\link{GdkFill}}]  the new fill mode.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
