\alias{gdkGCSetFill}
\name{gdkGCSetFill}
\title{gdkGCSetFill}
\description{Set the fill mode for a graphics context.}
\usage{gdkGCSetFill(object, fill)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkGC}}.}
\item{\verb{fill}}{the new fill mode.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
