\alias{gdkGCSetLineAttributes}
\name{gdkGCSetLineAttributes}
\title{gdkGCSetLineAttributes}
\description{Sets various attributes of how lines are drawn. See
the corresponding members of \code{\link{GdkGCValues}} for full
explanations of the arguments.}
\usage{gdkGCSetLineAttributes(object, line.width, line.style, cap.style, 
    join.style)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkGC}}.}
\item{\verb{line.width}}{the width of lines.}
\item{\verb{line.style}}{the dash-style for lines.}
\item{\verb{cap.style}}{the manner in which the ends of lines are drawn.}
\item{\verb{join.style}}{the in which lines are joined together.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
