\alias{gdkGCSetSubwindow}
\name{gdkGCSetSubwindow}
\title{gdkGCSetSubwindow}
\description{Sets how drawing with this GC on a window will affect child
windows of that window.}
\usage{gdkGCSetSubwindow(object, mode)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkGC}}.}
\item{\verb{mode}}{the subwindow mode.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
