\alias{gdkGCSetTile}
\name{gdkGCSetTile}
\title{gdkGCSetTile}
\description{Set a tile pixmap for a graphics context.
This will only be used if the fill mode
is \code{GDK_TILED}.}
\usage{gdkGCSetTile(object, tile)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkGC}}]   a \verb{\link{GdkGC}}.}
\item{\verb{tile}}{[\verb{\link{GdkPixmap}}]   the new tile pixmap.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
