\alias{gdkGCSetTile}
\name{gdkGCSetTile}
\title{gdkGCSetTile}
\description{Set a tile pixmap for a graphics context.
This will only be used if the fill mode
is \code{GDK_TILED}.}
\usage{gdkGCSetTile(object, tile)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkGC}}.}
\item{\verb{tile}}{the new tile pixmap.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
