\alias{gdkGCSetTsOrigin}
\name{gdkGCSetTsOrigin}
\title{gdkGCSetTsOrigin}
\description{Set the origin when using tiles or stipples with
the GC. The tile or stipple will be aligned such
that the upper left corner of the tile or stipple
will coincide with this point.}
\usage{gdkGCSetTsOrigin(object, x, y)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkGC}}]   a \verb{\link{GdkGC}}.}
\item{\verb{x}}{[integer]  the x-coordinate of the origin.}
\item{\verb{y}}{[integer]  the y-coordinate of the origin.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
