\alias{gdkImageGetPixel}
\name{gdkImageGetPixel}
\title{gdkImageGetPixel}
\description{Gets a pixel value at a specified position in a \code{\link{GdkImage}}.}
\usage{gdkImageGetPixel(object, x, y)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkImage}}.}
\item{\verb{x}}{the x coordinate of the pixel to get.}
\item{\verb{y}}{the y coordinate of the pixel to get.}
}
\value{[numeric] the pixel value at the given position.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
