\alias{gdkImagePutPixel}
\name{gdkImagePutPixel}
\title{gdkImagePutPixel}
\description{Sets a pixel in a \verb{\link{GdkImage}} to a given pixel value.}
\usage{gdkImagePutPixel(object, x, y, pixel)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkImage}}] a \verb{\link{GdkImage}}.}
\item{\verb{x}}{[integer] the x coordinate of the pixel to set.}
\item{\verb{y}}{[integer] the y coordinate of the pixel to set.}
\item{\verb{pixel}}{[numeric] the pixel value to set.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
