\alias{gdkImagePutPixel}
\name{gdkImagePutPixel}
\title{gdkImagePutPixel}
\description{Sets a pixel in a \code{\link{GdkImage}} to a given pixel value.}
\usage{gdkImagePutPixel(object, x, y, pixel)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkImage}}.}
\item{\verb{x}}{the x coordinate of the pixel to set.}
\item{\verb{y}}{the y coordinate of the pixel to set.}
\item{\verb{pixel}}{the pixel value to set.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
