\alias{gdkInputSetExtensionEvents}
\name{gdkInputSetExtensionEvents}
\title{gdkInputSetExtensionEvents}
\description{Turns extension events on or off for a particular window,
and specifies the event mask for extension events.}
\usage{gdkInputSetExtensionEvents(object, mask, mode)}
\arguments{
\item{\code{object}}{[\code{\link{GdkWindow}}] a \code{\link{GdkWindow}}.}
\item{\code{mask}}{[integer] the event mask}
\item{\code{mode}}{[\code{\link{GdkExtensionMode}}] the type of extension events that are desired.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
