\alias{gdkInputSetExtensionEvents}
\name{gdkInputSetExtensionEvents}
\title{gdkInputSetExtensionEvents}
\description{Turns extension events on or off for a particular window,
and specifies the event mask for extension events.}
\usage{gdkInputSetExtensionEvents(object, mask, mode)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkWindow}}.}
\item{\verb{mask}}{the event mask}
\item{\verb{mode}}{the type of extension events that are desired.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
