\alias{gdkKeymapGetDefault}
\name{gdkKeymapGetDefault}
\title{gdkKeymapGetDefault}
\description{Returns the \code{\link{GdkKeymap}} attached to the default display.}
\usage{gdkKeymapGetDefault()}
\value{[\code{\link{GdkKeymap}}] the \code{\link{GdkKeymap}} attached to the default display.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
