\alias{gdkKeymapGetDirection}
\name{gdkKeymapGetDirection}
\title{gdkKeymapGetDirection}
\description{Returns the direction of the keymap. }
\usage{gdkKeymapGetDirection(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkKeymap}}] a \code{\link{GdkKeymap}} or \code{NULL} to use the default keymap.
Returns: \code{PANGO_DIRECTION_LTR} or \code{PANGO_DIRECTION_RTL}.}}
\value{[\code{\link{PangoDirection}}] the direction of the keymap.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
