\alias{gdkKeymapGetForDisplay}
\name{gdkKeymapGetForDisplay}
\title{gdkKeymapGetForDisplay}
\description{Returns the \verb{\link{GdkKeymap}} attached to \code{display}.}
\usage{gdkKeymapGetForDisplay(display)}
\arguments{\item{\verb{display}}{[\verb{\link{GdkDisplay}}]  the \verb{\link{GdkDisplay}}.}}
\details{  Since  2.2}
\value{[\verb{\link{GdkKeymap}}]  the \verb{\link{GdkKeymap}} attached to \code{display}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
