\alias{gdkKeymapHaveBidiLayouts}
\name{gdkKeymapHaveBidiLayouts}
\title{gdkKeymapHaveBidiLayouts}
\description{Determines if keyboard layouts for both right-to-left and left-to-right
languages are in use.}
\usage{gdkKeymapHaveBidiLayouts(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkKeymap}} or \code{NULL} to use the default keymap}}
\details{Since 2.12}
\value{[logical] \code{TRUE} if there are layouts in both directions, \code{FALSE} otherwise}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
