\alias{gdkKeyvalFromName}
\name{gdkKeyvalFromName}
\title{gdkKeyvalFromName}
\description{Converts a key name to a key value.}
\usage{gdkKeyvalFromName(keyval.name)}
\arguments{\item{\verb{keyval.name}}{a key name.}}
\value{[numeric] the corresponding key value, or \code{GDK_VoidSymbol} if the key name is
not a valid key.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
