\alias{gdkPangoAttrEmbossColorNew}
\name{gdkPangoAttrEmbossColorNew}
\title{gdkPangoAttrEmbossColorNew}
\description{Creates a new attribute specifying the color to emboss text with.}
\usage{gdkPangoAttrEmbossColorNew(color)}
\arguments{\item{\verb{color}}{a GdkColor representing the color to emboss with}}
\details{Since 2.12}
\value{[\code{\link{PangoAttribute}}] new \code{\link{PangoAttribute}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
