\alias{gdkPangoAttrEmbossedNew}
\name{gdkPangoAttrEmbossedNew}
\title{gdkPangoAttrEmbossedNew}
\description{Creates a new attribute containing a embossed bitmap to be used when
rendering the text.}
\usage{gdkPangoAttrEmbossedNew(embossed)}
\arguments{\item{\code{embossed}}{[logical]  a bitmap to be set as embossed}}
\value{[\code{\link{PangoAttribute}}]  new \code{\link{PangoAttribute}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
