\alias{gdkPangoAttrStippleNew}
\name{gdkPangoAttrStippleNew}
\title{gdkPangoAttrStippleNew}
\description{Creates a new attribute containing a stipple bitmap to be used when
rendering the text.}
\usage{gdkPangoAttrStippleNew(stipple)}
\arguments{\item{\verb{stipple}}{a bitmap to be set as stipple}}
\value{[\code{\link{PangoAttribute}}] new \code{\link{PangoAttribute}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
