\alias{gdkPangoRendererSetDrawable}
\name{gdkPangoRendererSetDrawable}
\title{gdkPangoRendererSetDrawable}
\description{Sets the drawable the renderer draws to.}
\usage{gdkPangoRendererSetDrawable(object, drawable = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GdkPangoRenderer}}]  a \code{\link{GdkPangoRenderer}}}
\item{\code{drawable}}{[\code{\link{GdkDrawable}}]  the new target drawable, or \code{NULL}}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
