\alias{gdkPangoRendererSetDrawable}
\name{gdkPangoRendererSetDrawable}
\title{gdkPangoRendererSetDrawable}
\description{Sets the drawable the renderer draws to.}
\usage{gdkPangoRendererSetDrawable(object, drawable = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkPangoRenderer}}}
\item{\verb{drawable}}{the new target drawable, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
