\alias{gdkPixbufAnimationGetHeight}
\name{gdkPixbufAnimationGetHeight}
\title{gdkPixbufAnimationGetHeight}
\description{Queries the height of the bounding box of a pixbuf animation.}
\usage{gdkPixbufAnimationGetHeight(object)}
\arguments{\item{\verb{object}}{An animation.}}
\value{[integer] Height of the bounding box of the animation.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
