\alias{gdkPixbufFormatGetExtensions}
\name{gdkPixbufFormatGetExtensions}
\title{gdkPixbufFormatGetExtensions}
\description{Returns the filename extensions typically used for files in the 
given format.}
\usage{gdkPixbufFormatGetExtensions(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkPixbufFormat}}}}
\details{Since 2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
