\alias{gdkPixbufFormatGetMimeTypes}
\name{gdkPixbufFormatGetMimeTypes}
\title{gdkPixbufFormatGetMimeTypes}
\description{Returns the mime types supported by the format.}
\usage{gdkPixbufFormatGetMimeTypes(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkPixbufFormat}}]  a \verb{\link{GdkPixbufFormat}}}}
\details{  Since  2.2}
\value{[character] .}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
