\alias{gdkPixbufGetFormats}
\name{gdkPixbufGetFormats}
\title{gdkPixbufGetFormats}
\description{Obtains the available information about the image formats supported
by GdkPixbuf.}
\usage{gdkPixbufGetFormats()}
\details{Since 2.2}
\value{[list] A list of \code{\link{GdkPixbufFormat}}s describing the supported
image formats.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
