\alias{gdkPixbufGetHasAlpha}
\name{gdkPixbufGetHasAlpha}
\title{gdkPixbufGetHasAlpha}
\description{Queries whether a pixbuf has an alpha channel (opacity information).}
\usage{gdkPixbufGetHasAlpha(object)}
\arguments{\item{\verb{object}}{A pixbuf.}}
\value{[logical] \code{TRUE} if it has an alpha channel, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
