\alias{gdkPixbufGetWidth}
\name{gdkPixbufGetWidth}
\title{gdkPixbufGetWidth}
\description{Queries the width of a pixbuf.}
\usage{gdkPixbufGetWidth(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkPixbuf}}]  A pixbuf.}}
\value{[integer]  Width in pixels.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
