\alias{gdkPixbufLoaderGetFormat}
\name{gdkPixbufLoaderGetFormat}
\title{gdkPixbufLoaderGetFormat}
\description{Obtains the available information about the format of the 
currently loading image file.}
\usage{gdkPixbufLoaderGetFormat(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkPixbufLoader}}]  A pixbuf loader.}}
\details{  Since  2.2}
\value{[\verb{\link{GdkPixbufFormat}}]  A \verb{\link{GdkPixbufFormat}} or \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
