\alias{gdkPixbufRotateSimple}
\name{gdkPixbufRotateSimple}
\title{gdkPixbufRotateSimple}
\description{Rotates a pixbuf by a multiple of 90 degrees, and returns the
result in a new pixbuf.}
\usage{gdkPixbufRotateSimple(object, angle)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkPixbuf}}]  a \verb{\link{GdkPixbuf}}}
\item{\verb{angle}}{[\code{\link{GdkPixbufRotation}}]  the angle to rotate by}
}
\details{  Since  2.6}
\value{[\verb{\link{GdkPixbuf}}]  the new \verb{\link{GdkPixbuf}}, or \code{NULL} if not enough memory could be
allocated for it.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
