\alias{gdkPixbufSimpleAnimAddFrame}
\name{gdkPixbufSimpleAnimAddFrame}
\title{gdkPixbufSimpleAnimAddFrame}
\description{Adds a new frame to \code{animation}. The \code{pixbuf} must
have the dimensions specified when the animation 
was constructed.}
\usage{gdkPixbufSimpleAnimAddFrame(object, pixbuf)}
\arguments{
\item{\code{object}}{[\code{\link{GdkPixbufSimpleAnim}}]  a \code{\link{GdkPixbufSimpleAnim}}}
\item{\code{pixbuf}}{[\code{\link{GdkPixbuf}}]  the pixbuf to add }
}
\details{  Since  2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
