\alias{gdkPixbufSimpleAnimNew}
\name{gdkPixbufSimpleAnimNew}
\title{gdkPixbufSimpleAnimNew}
\description{Creates a new, empty animation.}
\usage{gdkPixbufSimpleAnimNew(width, height, rate)}
\arguments{
\item{\code{width}}{[integer]  the width of the animation}
\item{\code{height}}{[integer]  the height of the animation}
\item{\code{rate}}{[numeric]  the speed of the animation, in frames per second}
}
\details{  Since  2.8}
\value{[\code{\link{GdkPixbufSimpleAnim}}]  a newly allocated \code{\link{GdkPixbufSimpleAnim}} }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
