\alias{gdkPointerUngrab}
\name{gdkPointerUngrab}
\title{gdkPointerUngrab}
\description{Ungrabs the pointer on the default display, if it is grabbed by this 
application.}
\usage{gdkPointerUngrab(time = "GDK_CURRENT_TIME")}
\arguments{\item{\verb{time}}{a timestamp from a \code{\link{GdkEvent}}, or \code{GDK_CURRENT_TIME} if no
timestamp is available.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
