\alias{gdkPropertyDelete}
\name{gdkPropertyDelete}
\title{gdkPropertyDelete}
\description{Deletes a property from a window.}
\usage{gdkPropertyDelete(object, property)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkWindow}}.}
\item{\verb{property}}{the property to delete.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
