\alias{gdkRegionCopy}
\name{gdkRegionCopy}
\title{gdkRegionCopy}
\description{Copies \code{region}, creating an identical new region.}
\usage{gdkRegionCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkRegion}}]  a \verb{\link{GdkRegion}}}}
\value{[\verb{\link{GdkRegion}}]  a new region identical to \code{region}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
