\alias{gdkRegionEmpty}
\name{gdkRegionEmpty}
\title{gdkRegionEmpty}
\description{Returns \code{TRUE} if the \code{\link{GdkRegion}} is empty.}
\usage{gdkRegionEmpty(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkRegion}}] a \code{\link{GdkRegion}}.}}
\value{[logical] \code{TRUE} if \code{region} is empty.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
