\alias{gdkRegionOffset}
\name{gdkRegionOffset}
\title{gdkRegionOffset}
\description{Moves a region the specified distance.}
\usage{gdkRegionOffset(object, dx, dy)}
\arguments{
\item{\code{object}}{[\code{\link{GdkRegion}}] a \code{\link{GdkRegion}}.}
\item{\code{dx}}{[integer] the distance to move the region horizontally.}
\item{\code{dy}}{[integer] the distance to move the region vertically.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
