\alias{gdkRegionOffset}
\name{gdkRegionOffset}
\title{gdkRegionOffset}
\description{Moves a region the specified distance.}
\usage{gdkRegionOffset(object, dx, dy)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkRegion}}}
\item{\verb{dx}}{the distance to move the region horizontally}
\item{\verb{dy}}{the distance to move the region vertically}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
