\alias{gdkRegionPointIn}
\name{gdkRegionPointIn}
\title{gdkRegionPointIn}
\description{Finds out if a point is in a region.}
\usage{gdkRegionPointIn(object, x, y)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkRegion}}]  a \verb{\link{GdkRegion}}}
\item{\verb{x}}{[integer]  the x coordinate of a point}
\item{\verb{y}}{[integer]  the y coordinate of a point}
}
\value{[logical] \code{TRUE} if the point is in \code{region}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
