\alias{gdkRegionPointIn}
\name{gdkRegionPointIn}
\title{gdkRegionPointIn}
\description{Finds out if a point is in a region.}
\usage{gdkRegionPointIn(object, x, y)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkRegion}}}
\item{\verb{x}}{the x coordinate of a point}
\item{\verb{y}}{the y coordinate of a point}
}
\value{[logical] \code{TRUE} if the point is in \code{region}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
