\alias{gdkRegionRectangle}
\name{gdkRegionRectangle}
\title{gdkRegionRectangle}
\description{Creates a new region containing the area \code{rectangle}.}
\usage{gdkRegionRectangle(rectangle)}
\arguments{\item{\code{rectangle}}{[\code{\link{GdkRectangle}}]  a \code{\link{GdkRectangle}}}}
\value{[\code{\link{GdkRegion}}]  a new region}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
