\alias{gdkRegionShrink}
\name{gdkRegionShrink}
\title{gdkRegionShrink}
\description{Resizes a region by the specified amount.
Positive values shrink the region. Negative values expand it.}
\usage{gdkRegionShrink(object, dx, dy)}
\arguments{
\item{\code{object}}{[\code{\link{GdkRegion}}]  a \code{\link{GdkRegion}}}
\item{\code{dx}}{[integer]  the number of pixels to shrink the region horizontally}
\item{\code{dy}}{[integer]  the number of pixels to shrink the region vertically}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
