\alias{gdkRegionShrink}
\name{gdkRegionShrink}
\title{gdkRegionShrink}
\description{Resizes a region by the specified amount.
Positive values shrink the region. Negative values expand it.}
\usage{gdkRegionShrink(object, dx, dy)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkRegion}}}
\item{\verb{dx}}{the number of pixels to shrink the region horizontally}
\item{\verb{dy}}{the number of pixels to shrink the region vertically}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
