\alias{gdkScreenGetDefault}
\name{gdkScreenGetDefault}
\title{gdkScreenGetDefault}
\description{Gets the default screen for the default display. (See
\code{\link{gdkDisplayGetDefault}}).}
\usage{gdkScreenGetDefault()}
\details{Since 2.2}
\value{[\code{\link{GdkScreen}}] a \code{\link{GdkScreen}}, or \code{NULL} if there is no default display. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
