\alias{gdkScreenGetDefaultColormap}
\name{gdkScreenGetDefaultColormap}
\title{gdkScreenGetDefaultColormap}
\description{Gets the default colormap for \code{screen}.}
\usage{gdkScreenGetDefaultColormap(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkScreen}}]  a \verb{\link{GdkScreen}}}}
\details{  Since  2.2}
\value{[\verb{\link{GdkColormap}}]  the default \verb{\link{GdkColormap}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
