\alias{gdkScreenGetDefaultColormap}
\name{gdkScreenGetDefaultColormap}
\title{gdkScreenGetDefaultColormap}
\description{Gets the default colormap for \code{screen}.}
\usage{gdkScreenGetDefaultColormap(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkScreen}}}}
\details{Since 2.2}
\value{[\code{\link{GdkColormap}}] the default \code{\link{GdkColormap}}. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
