\alias{gdkScreenGetDisplay}
\name{gdkScreenGetDisplay}
\title{gdkScreenGetDisplay}
\description{Gets the display to which the \code{screen} belongs.}
\usage{gdkScreenGetDisplay(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkScreen}}}}
\details{Since 2.2}
\value{[\code{\link{GdkDisplay}}] the display to which \code{screen} belongs}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
