\alias{gdkScreenGetFontOptions}
\name{gdkScreenGetFontOptions}
\title{gdkScreenGetFontOptions}
\description{Gets any options previously set with \code{\link{gdkScreenSetFontOptions}}.}
\usage{gdkScreenGetFontOptions(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkScreen}}}}
\details{Since 2.10}
\value{[\code{\link{CairoFontOptions}}] the current font options, or \code{NULL} if no default
font options have been set.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
