\alias{gdkScreenGetRootWindow}
\name{gdkScreenGetRootWindow}
\title{gdkScreenGetRootWindow}
\description{Gets the root window of \code{screen}.}
\usage{gdkScreenGetRootWindow(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkScreen}}]  a \code{\link{GdkScreen}}}}
\details{  Since  2.2}
\value{[\code{\link{GdkWindow}}]  the root window}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
