\alias{gdkScreenGetRootWindow}
\name{gdkScreenGetRootWindow}
\title{gdkScreenGetRootWindow}
\description{Gets the root window of \code{screen}.}
\usage{gdkScreenGetRootWindow(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkScreen}}}}
\details{Since 2.2}
\value{[\code{\link{GdkWindow}}] the root window. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
