\alias{gdkScreenGetSystemColormap}
\name{gdkScreenGetSystemColormap}
\title{gdkScreenGetSystemColormap}
\description{Gets the system's default colormap for \code{screen}}
\usage{gdkScreenGetSystemColormap(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkScreen}}]  a \verb{\link{GdkScreen}}}}
\details{  Since  2.2}
\value{[\verb{\link{GdkColormap}}]  the default colormap for \code{screen}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
