\alias{gdkScreenGetSystemColormap}
\name{gdkScreenGetSystemColormap}
\title{gdkScreenGetSystemColormap}
\description{Gets the system's default colormap for \code{screen}}
\usage{gdkScreenGetSystemColormap(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkScreen}}}}
\details{Since 2.2}
\value{[\code{\link{GdkColormap}}] the default colormap for \code{screen}. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
