\alias{gdkScreenGetWidth}
\name{gdkScreenGetWidth}
\title{gdkScreenGetWidth}
\description{Gets the width of \code{screen} in pixels}
\usage{gdkScreenGetWidth(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkScreen}}]  a \code{\link{GdkScreen}}}}
\details{  Since  2.2}
\value{[integer]  the width of \code{screen} in pixels.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
