\alias{gdkScreenMakeDisplayName}
\name{gdkScreenMakeDisplayName}
\title{gdkScreenMakeDisplayName}
\description{Determines the name to pass to \code{\link{gdkDisplayOpen}} to get
a \code{\link{GdkDisplay}} with this screen as the default screen.}
\usage{gdkScreenMakeDisplayName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkScreen}}}}
\details{Since 2.2}
\value{[character] a newly allocated string,}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
