\alias{gdkScreenSetDefaultColormap}
\name{gdkScreenSetDefaultColormap}
\title{gdkScreenSetDefaultColormap}
\description{Sets the default \code{colormap} for \code{screen}.}
\usage{gdkScreenSetDefaultColormap(object, colormap)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkScreen}}}
\item{\verb{colormap}}{a \code{\link{GdkColormap}}}
}
\details{Since 2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
