\alias{gdkTextWidthWc}
\name{gdkTextWidthWc}
\title{gdkTextWidthWc}
\description{
Determines the width of a given wide-character string.
\strong{WARNING: \code{gdk_text_width_wc} is deprecated and should not be used in newly-written code.}
}
\usage{gdkTextWidthWc(object, text)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkFont}}}
\item{\verb{text}}{the text to measure.}
}
\value{[integer] the width of the string in pixels.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
