\alias{gdkVisualGetBest}
\name{gdkVisualGetBest}
\title{gdkVisualGetBest}
\description{Get the visual with the most available colors for the default
GDK screen.}
\usage{gdkVisualGetBest()}
\value{[\verb{\link{GdkVisual}}]  best visual}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
