\alias{gdkVisualGetBest}
\name{gdkVisualGetBest}
\title{gdkVisualGetBest}
\description{Get the visual with the most available colors for the default
GDK screen. The return value should not be freed.}
\usage{gdkVisualGetBest()}
\value{[\code{\link{GdkVisual}}] best visual. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
