\alias{gdkVisualGetScreen}
\name{gdkVisualGetScreen}
\title{gdkVisualGetScreen}
\description{Gets the screen to which this visual belongs}
\usage{gdkVisualGetScreen(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkVisual}}}}
\details{Since 2.2}
\value{[\code{\link{GdkScreen}}] the screen to which this visual belongs. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
