\alias{gdkVisualGetSystem}
\name{gdkVisualGetSystem}
\title{gdkVisualGetSystem}
\description{Get the system's default visual for the default GDK screen.
This is the visual for the root window of the display.
The return value should not be freed.}
\usage{gdkVisualGetSystem()}
\value{[\code{\link{GdkVisual}}] system visual. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
