\alias{gdkWindowBeginPaintRect}
\name{gdkWindowBeginPaintRect}
\title{gdkWindowBeginPaintRect}
\description{A convenience wrapper around \code{\link{gdkWindowBeginPaintRegion}} which
creates a rectangular region for you. See
\code{\link{gdkWindowBeginPaintRegion}} for details.}
\usage{gdkWindowBeginPaintRect(object, rectangle)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkWindow}}}
\item{\verb{rectangle}}{rectangle you intend to draw to}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
