\alias{gdkWindowClear}
\name{gdkWindowClear}
\title{gdkWindowClear}
\description{Clears an entire \code{window} to the background color or background pixmap.}
\usage{gdkWindowClear(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkWindow}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
