\alias{gdkWindowClearArea}
\name{gdkWindowClearArea}
\title{gdkWindowClearArea}
\description{Clears an area of \code{window} to the background color or background pixmap.}
\usage{gdkWindowClearArea(object, x, y, width, height)}
\arguments{
\item{\verb{object}}{[\verb{\link{GdkWindow}}]  a \verb{\link{GdkWindow}}}
\item{\verb{x}}{[integer]  x coordinate of rectangle to clear}
\item{\verb{y}}{[integer]  y coordinate of rectangle to clear}
\item{\verb{width}}{[integer]  width of rectangle to clear}
\item{\verb{height}}{[integer]  height of rectangle to clear}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
