\alias{gdkWindowClearArea}
\name{gdkWindowClearArea}
\title{gdkWindowClearArea}
\description{Clears an area of \code{window} to the background color or background pixmap.}
\usage{gdkWindowClearArea(object, x, y, width, height)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkWindow}}}
\item{\verb{x}}{x coordinate of rectangle to clear}
\item{\verb{y}}{y coordinate of rectangle to clear}
\item{\verb{width}}{width of rectangle to clear}
\item{\verb{height}}{height of rectangle to clear}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
