\alias{gdkWindowGetGroup}
\name{gdkWindowGetGroup}
\title{gdkWindowGetGroup}
\description{Returns the group leader window for \code{window}. See \code{\link{gdkWindowSetGroup}}.}
\usage{gdkWindowGetGroup(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkWindow}}]  a toplevel \verb{\link{GdkWindow}}}}
\details{  Since  2.4}
\value{[\verb{\link{GdkWindow}}]  the group leader window for \code{window}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
