\alias{gdkWindowGetState}
\name{gdkWindowGetState}
\title{gdkWindowGetState}
\description{Gets the bitwise OR of the currently active window state flags,
from the \code{\link{GdkWindowState}} enumeration.}
\usage{gdkWindowGetState(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkWindow}}}}
\value{[\code{\link{GdkWindowState}}] window state bitfield}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
