\alias{gdkWindowGetToplevel}
\name{gdkWindowGetToplevel}
\title{gdkWindowGetToplevel}
\description{Gets the toplevel window that's an ancestor of \code{window}.}
\usage{gdkWindowGetToplevel(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkWindow}}]  a \verb{\link{GdkWindow}}}}
\value{[\verb{\link{GdkWindow}}]  the toplevel window containing \code{window}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
