\alias{gdkWindowGetToplevels}
\name{gdkWindowGetToplevels}
\title{gdkWindowGetToplevels}
\description{Obtains a list of all toplevel windows known to GDK on the default
screen (see \code{gdk\Window\Get\Toplevels\For\Screen()}).
A toplevel window is a child of the root window (see
\code{\link{gdkGetDefaultRootWindow}}).}
\usage{gdkWindowGetToplevels()}
\details{  .}
\value{[list]  list of toplevel windows, free with \code{g\List\Free()}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
