\alias{gdkWindowGetTypeHint}
\name{gdkWindowGetTypeHint}
\title{gdkWindowGetTypeHint}
\description{This function returns the type hint set for a window.}
\usage{gdkWindowGetTypeHint(object)}
\arguments{\item{\verb{object}}{A toplevel \code{\link{GdkWindow}}}}
\details{Since 2.10}
\value{[\code{\link{GdkWindowTypeHint}}] The type hint set for \code{window}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
