\alias{gdkWindowGetWindowType}
\name{gdkWindowGetWindowType}
\title{gdkWindowGetWindowType}
\description{Gets the type of the window. See \code{\link{GdkWindowType}}.}
\usage{gdkWindowGetWindowType(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkWindow}}}}
\value{[\code{\link{GdkWindowType}}] type of window}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
