\alias{gdkWindowPeekChildren}
\name{gdkWindowPeekChildren}
\title{gdkWindowPeekChildren}
\description{Like \code{\link{gdkWindowGetChildren}}, but does not copy the list of
children.}
\usage{gdkWindowPeekChildren(object)}
\arguments{\item{\code{object}}{[\code{\link{GdkWindow}}]  a \code{\link{GdkWindow}}}}
\value{[list]  a reference to the list of child windows in \code{window}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
