\alias{gdkWindowRegisterDnd}
\name{gdkWindowRegisterDnd}
\title{gdkWindowRegisterDnd}
\description{Registers a window as a potential drop destination.}
\usage{gdkWindowRegisterDnd(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GdkWindow}}] a \verb{\link{GdkWindow}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
