\alias{gdkWindowRemoveFilter}
\name{gdkWindowRemoveFilter}
\title{gdkWindowRemoveFilter}
\description{Remove a filter previously added with \code{\link{gdkWindowAddFilter}}.}
\usage{gdkWindowRemoveFilter(object, fun, data)}
\arguments{
\item{\code{object}}{[\code{\link{GdkWindow}}]  a \code{\link{GdkWindow}}}
\item{\code{data}}{[R object]  user data for previously-added filter function}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
