\alias{gdkWindowRemoveFilter}
\name{gdkWindowRemoveFilter}
\title{gdkWindowRemoveFilter}
\description{Remove a filter previously added with \code{\link{gdkWindowAddFilter}}.}
\usage{gdkWindowRemoveFilter(object, fun, data)}
\arguments{
\item{\verb{object}}{a \code{\link{GdkWindow}}}
\item{\verb{data}}{user data for previously-added filter function}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
