\alias{gdkWindowSetStartupId}
\name{gdkWindowSetStartupId}
\title{gdkWindowSetStartupId}
\description{When using GTK+, typically you should use \code{\link{gtkWindowSetStartupId}}
instead of this low-level function.}
\usage{gdkWindowSetStartupId(object, startup.id)}
\arguments{
\item{\verb{object}}{a toplevel \code{\link{GdkWindow}}}
\item{\verb{startup.id}}{a string with startup-notification identifier}
}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
