\alias{gdkWindowSetUrgencyHint}
\name{gdkWindowSetUrgencyHint}
\title{gdkWindowSetUrgencyHint}
\description{Toggles whether a window needs the user's
urgent attention.}
\usage{gdkWindowSetUrgencyHint(object, urgent)}
\arguments{
\item{\verb{object}}{a toplevel \code{\link{GdkWindow}}}
\item{\verb{urgent}}{\code{TRUE} if the window is urgent}
}
\details{Since 2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
