\alias{gdkWindowThawUpdates}
\name{gdkWindowThawUpdates}
\title{gdkWindowThawUpdates}
\description{Thaws a window frozen with \code{\link{gdkWindowFreezeUpdates}}.}
\usage{gdkWindowThawUpdates(object)}
\arguments{\item{\verb{object}}{a \code{\link{GdkWindow}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
