\alias{gdkWindowUnstick}
\name{gdkWindowUnstick}
\title{gdkWindowUnstick}
\description{Reverse operation for \code{\link{gdkWindowStick}}; see \code{\link{gdkWindowStick}},
and \code{\link{gtkWindowUnstick}}.}
\usage{gdkWindowUnstick(object)}
\arguments{\item{\verb{object}}{a toplevel \code{\link{GdkWindow}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
