\alias{gladeGetWidgetTree}
\name{gladeGetWidgetTree}
\title{gladeGetWidgetTree}
\description{This function is used to get the GladeXML object that built this widget.}
\usage{gladeGetWidgetTree(widget)}
\arguments{\item{\verb{widget}}{[\code{\link{GtkWidget}}]  the widget}}
\value{[\code{\link{GladeXML}}]  the GladeXML object that built this widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
