\alias{gtkAboutDialogGetCopyright}
\name{gtkAboutDialogGetCopyright}
\title{gtkAboutDialogGetCopyright}
\description{Returns the copyright string.}
\usage{gtkAboutDialogGetCopyright(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}}
\details{Since 2.6}
\value{[character] The copyright string.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
