\alias{gtkAboutDialogGetDocumenters}
\name{gtkAboutDialogGetDocumenters}
\title{gtkAboutDialogGetDocumenters}
\description{Returns the string which are displayed in the documenters
tab of the secondary credits dialog.}
\usage{gtkAboutDialogGetDocumenters(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}}
\details{Since 2.6}
\value{[character] A string list containing
the documenters.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
