\alias{gtkAboutDialogGetLicense}
\name{gtkAboutDialogGetLicense}
\title{gtkAboutDialogGetLicense}
\description{Returns the license information.}
\usage{gtkAboutDialogGetLicense(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}}
\details{Since 2.6}
\value{[character] The license information.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
