\alias{gtkAboutDialogGetLogo}
\name{gtkAboutDialogGetLogo}
\title{gtkAboutDialogGetLogo}
\description{Returns the pixbuf displayed as logo in the about dialog.}
\usage{gtkAboutDialogGetLogo(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}}
\details{Since 2.6}
\value{[\code{\link{GdkPixbuf}}] the pixbuf displayed as logo. If you want to keep a reference
to it,}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
