\alias{gtkAboutDialogGetName}
\name{gtkAboutDialogGetName}
\title{gtkAboutDialogGetName}
\description{
Returns the program name displayed in the about dialog.
\strong{WARNING: \code{gtk_about_dialog_get_name} is deprecated and should not be used in newly-written code.}
}
\usage{gtkAboutDialogGetName(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkAboutDialog}}]  a \code{\link{GtkAboutDialog}}}}
\details{  Since  2.6

@Deprecated: 2.12: Use gtk\_about\_dialog\_get\_program\_name() instead.}
\value{[character]  The program name.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
