\alias{gtkAboutDialogGetName}
\name{gtkAboutDialogGetName}
\title{gtkAboutDialogGetName}
\description{
Returns the program name displayed in the about dialog.
\strong{WARNING: \code{gtk_about_dialog_get_name} has been deprecated since version 2.12 and should not be used in newly-written code. Use \code{\link{gtkAboutDialogGetProgramName}} instead.}
}
\usage{gtkAboutDialogGetName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}}
\details{Since 2.6}
\value{[character] The program name.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
