\alias{gtkAboutDialogGetWebsite}
\name{gtkAboutDialogGetWebsite}
\title{gtkAboutDialogGetWebsite}
\description{Returns the website URL.}
\usage{gtkAboutDialogGetWebsite(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkAboutDialog}}]  a \verb{\link{GtkAboutDialog}}}}
\details{  Since  2.6}
\value{[character]  The website URL.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
