\alias{gtkAboutDialogGetWebsiteLabel}
\name{gtkAboutDialogGetWebsiteLabel}
\title{gtkAboutDialogGetWebsiteLabel}
\description{Returns the label used for the website link.}
\usage{gtkAboutDialogGetWebsiteLabel(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}}
\details{Since 2.6}
\value{[character] The label used for the website link.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
