\alias{gtkAboutDialogNew}
\name{gtkAboutDialogNew}
\title{gtkAboutDialogNew}
\description{Creates a new \verb{\link{GtkAboutDialog}}.}
\usage{gtkAboutDialogNew(show = TRUE)}
\details{  Since  2.6}
\value{[\verb{\link{GtkWidget}}]  a newly created \verb{\link{GtkAboutDialog}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
