\alias{gtkAboutDialogNew}
\name{gtkAboutDialogNew}
\title{gtkAboutDialogNew}
\description{Creates a new \code{\link{GtkAboutDialog}}.}
\usage{gtkAboutDialogNew(show = TRUE)}
\details{Since 2.6}
\value{[\code{\link{GtkWidget}}] a newly created \code{\link{GtkAboutDialog}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
