\alias{gtkAboutDialogSetAuthors}
\name{gtkAboutDialogSetAuthors}
\title{gtkAboutDialogSetAuthors}
\description{Sets the strings which are displayed in the authors tab
of the secondary credits dialog.}
\usage{gtkAboutDialogSetAuthors(object, authors)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}
\item{\verb{authors}}{a list of strings}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
