\alias{gtkAboutDialogSetComments}
\name{gtkAboutDialogSetComments}
\title{gtkAboutDialogSetComments}
\description{Sets the comments string to display in the about dialog.
This should be a short string of one or two lines.}
\usage{gtkAboutDialogSetComments(object, comments = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}
\item{\verb{comments}}{a comments string. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
