\alias{gtkAboutDialogSetCopyright}
\name{gtkAboutDialogSetCopyright}
\title{gtkAboutDialogSetCopyright}
\description{Sets the copyright string to display in the about dialog.
This should be a short string of one or two lines.}
\usage{gtkAboutDialogSetCopyright(object, copyright = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}
\item{\verb{copyright}}{(allow-none) the copyright string}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
