\alias{gtkAboutDialogSetDocumenters}
\name{gtkAboutDialogSetDocumenters}
\title{gtkAboutDialogSetDocumenters}
\description{Sets the strings which are displayed in the documenters tab
of the secondary credits dialog.}
\usage{gtkAboutDialogSetDocumenters(object, documenters)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}
\item{\verb{documenters}}{a list of strings}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
