\alias{gtkAboutDialogSetLicense}
\name{gtkAboutDialogSetLicense}
\title{gtkAboutDialogSetLicense}
\description{Sets the license information to be displayed in the secondary
license dialog. If \code{license} is \code{NULL}, the license button is
hidden.}
\usage{gtkAboutDialogSetLicense(object, license = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkAboutDialog}}]  a \verb{\link{GtkAboutDialog}}}
\item{\verb{license}}{[character]  the license information or \code{NULL}}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
