\alias{gtkAboutDialogSetLogo}
\name{gtkAboutDialogSetLogo}
\title{gtkAboutDialogSetLogo}
\description{Sets the pixbuf to be displayed as logo in the about dialog.
If it is \code{NULL}, the default window icon set with
\code{\link{gtkWindowSetDefaultIcon}} will be used.}
\usage{gtkAboutDialogSetLogo(object, logo = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}
\item{\verb{logo}}{a \code{\link{GdkPixbuf}}, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
