\alias{gtkAboutDialogSetName}
\name{gtkAboutDialogSetName}
\title{gtkAboutDialogSetName}
\description{Sets the name to display in the about dialog. 
If this is not set, it defaults to \code{g\Get\Application\Name()}.}
\usage{gtkAboutDialogSetName(object, name = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAboutDialog}}]  a \code{\link{GtkAboutDialog}}}
\item{\code{name}}{[character]  the program name}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
