\alias{gtkAboutDialogSetProgramName}
\name{gtkAboutDialogSetProgramName}
\title{gtkAboutDialogSetProgramName}
\description{Sets the name to display in the about dialog.
If this is not set, it defaults to \code{gGetApplicationName()}.}
\usage{gtkAboutDialogSetProgramName(object, name)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}
\item{\verb{name}}{the program name}
}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
