\alias{gtkAboutDialogSetVersion}
\name{gtkAboutDialogSetVersion}
\title{gtkAboutDialogSetVersion}
\description{Sets the version string to display in the about dialog.}
\usage{gtkAboutDialogSetVersion(object, version = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAboutDialog}}]  a \code{\link{GtkAboutDialog}}}
\item{\code{version}}{[character]  the version string }
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
