\alias{gtkAboutDialogSetVersion}
\name{gtkAboutDialogSetVersion}
\title{gtkAboutDialogSetVersion}
\description{Sets the version string to display in the about dialog.}
\usage{gtkAboutDialogSetVersion(object, version = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}
\item{\verb{version}}{the version string. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
