\alias{gtkAboutDialogSetWebsite}
\name{gtkAboutDialogSetWebsite}
\title{gtkAboutDialogSetWebsite}
\description{Sets the URL to use for the website link.}
\usage{gtkAboutDialogSetWebsite(object, website = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}
\item{\verb{website}}{a URL string starting with "http://". \emph{[  \acronym{allow-none}  ]}}
}
\details{Note that that the hook functions need to be set up
before calling this function.
  Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
