\alias{gtkAboutDialogSetWebsiteLabel}
\name{gtkAboutDialogSetWebsiteLabel}
\title{gtkAboutDialogSetWebsiteLabel}
\description{Sets the label to be used for the website link.
It defaults to the website URL.}
\usage{gtkAboutDialogSetWebsiteLabel(object, website.label = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}
\item{\verb{website.label}}{the label used for the website link}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
