\alias{gtkAboutDialogSetWrapLicense}
\name{gtkAboutDialogSetWrapLicense}
\title{gtkAboutDialogSetWrapLicense}
\description{Sets whether the license text in \code{about} is
automatically wrapped.}
\usage{gtkAboutDialogSetWrapLicense(object, wrap.license)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAboutDialog}}}
\item{\verb{wrap.license}}{whether to wrap the license}
}
\details{Since 2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
