\alias{gtkAccelGroupActivate}
\name{gtkAccelGroupActivate}
\title{gtkAccelGroupActivate}
\usage{gtkAccelGroupActivate(object, accel.quark, acceleratable, accel.key, accel.mods)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAccelGroup}}] }
\item{\code{accel.quark}}{[numeric] }
\item{\code{acceleratable}}{[\code{\link{GObject}}] }
\item{\code{accel.key}}{[numeric] }
\item{\code{accel.mods}}{[\code{\link{GdkModifierType}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
