\alias{gtkAccelGroupFromAccelClosure}
\name{gtkAccelGroupFromAccelClosure}
\title{gtkAccelGroupFromAccelClosure}
\description{Finds the \verb{\link{GtkAccelGroup}} to which \code{closure} is connected; 
see \code{\link{gtkAccelGroupConnect}}.}
\usage{gtkAccelGroupFromAccelClosure(closure)}
\arguments{\item{\verb{closure}}{[\verb{\link{GClosure}}]  a \verb{\link{GClosure}}}}
\value{[\verb{\link{GtkAccelGroup}}]  the \verb{\link{GtkAccelGroup}} to which \code{closure} is connected, or \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
