\alias{gtkAccelGroupNew}
\name{gtkAccelGroupNew}
\title{gtkAccelGroupNew}
\description{Creates a new \verb{\link{GtkAccelGroup}}.}
\usage{gtkAccelGroupNew()}
\value{[\verb{\link{GtkAccelGroup}}]  a new \verb{\link{GtkAccelGroup}} object}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
