\alias{gtkAccelGroupsFromObject}
\name{gtkAccelGroupsFromObject}
\title{gtkAccelGroupsFromObject}
\description{Gets a list of all accel groups which are attached to \code{object}.}
\usage{gtkAccelGroupsFromObject(object)}
\arguments{\item{\verb{object}}{a \code{\link{GObject}}, usually a \code{\link{GtkWindow}}}}
\value{[list] a list of all accel groups which are attached to \code{object}. \emph{[  \acronym{element-type}   GtkAccelGroup][  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
