\alias{gtkAccelGroupUnlock}
\name{gtkAccelGroupUnlock}
\title{gtkAccelGroupUnlock}
\description{Undoes the last call to \code{\link{gtkAccelGroupLock}} on this \code{accel.group}.}
\usage{gtkAccelGroupUnlock(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAccelGroup}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
