\alias{gtkAccelLabelGetAccelWidget}
\name{gtkAccelLabelGetAccelWidget}
\title{gtkAccelLabelGetAccelWidget}
\description{Fetches the widget monitored by this accelerator label. See
\code{\link{gtkAccelLabelSetAccelWidget}}.}
\usage{gtkAccelLabelGetAccelWidget(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAccelLabel}}}}
\value{[\code{\link{GtkWidget}}] the object monitored by the accelerator label, or \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
