\alias{gtkAccelLabelNew}
\name{gtkAccelLabelNew}
\title{gtkAccelLabelNew}
\description{Creates a new \code{\link{GtkAccelLabel}}.}
\usage{gtkAccelLabelNew(string = NULL, show = TRUE)}
\arguments{\item{\verb{string}}{the label string. Must be non-\code{NULL}.}}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkAccelLabel}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
