\alias{gtkAccelLabelSetAccelClosure}
\name{gtkAccelLabelSetAccelClosure}
\title{gtkAccelLabelSetAccelClosure}
\description{Sets the closure to be monitored by this accelerator label. The closure
must be connected to an accelerator group; see \code{\link{gtkAccelGroupConnect}}.}
\usage{gtkAccelLabelSetAccelClosure(object, accel.closure)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAccelLabel}}}
\item{\verb{accel.closure}}{the closure to monitor for accelerator changes.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
