\alias{gtkAccelLabelSetAccelWidget}
\name{gtkAccelLabelSetAccelWidget}
\title{gtkAccelLabelSetAccelWidget}
\description{Sets the widget to be monitored by this accelerator label.}
\usage{gtkAccelLabelSetAccelWidget(object, accel.widget)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAccelLabel}}]  a \code{\link{GtkAccelLabel}}}
\item{\code{accel.widget}}{[\code{\link{GtkWidget}}]  the widget to be monitored.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
