\alias{gtkAccelMapLoadScanner}
\name{gtkAccelMapLoadScanner}
\title{gtkAccelMapLoadScanner}
\description{\verb{GScanner} variant of \code{\link{gtkAccelMapLoad}}.}
\usage{gtkAccelMapLoadScanner(scanner)}
\arguments{\item{\verb{scanner}}{a \verb{GScanner} which has already been provided with an input file}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
