\alias{gtkAccelMapSave}
\name{gtkAccelMapSave}
\title{gtkAccelMapSave}
\description{Saves current accelerator specifications (accelerator path, key
and modifiers) to \code{file.name}.
The file is written in a format suitable to be read back in by
\code{\link{gtkAccelMapLoad}}.}
\usage{gtkAccelMapSave(file.name)}
\arguments{\item{\verb{file.name}}{the name of the file to contain accelerator specifications,
  in the GLib file name encoding}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
