\alias{gtkAccelMapSaveFd}
\name{gtkAccelMapSaveFd}
\title{gtkAccelMapSaveFd}
\description{Filedescriptor variant of \code{\link{gtkAccelMapSave}}.}
\usage{gtkAccelMapSaveFd(fd)}
\arguments{\item{\code{fd}}{[integer]  a valid writable file descriptor}}
\details{Note that the file descriptor will not be closed by this function.  }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
