\alias{GtkAccessible}
\name{GtkAccessible}
\title{GtkAccessible}
\description{Accessibility support for widgets}
\usage{\link{gtkAccessibleConnectWidgetDestroyed}(object)}
\section{Hierarchy}{\preformatted{  \link{GObject}
   +----\link{AtkObject}
         +----GtkAccessible}}
\section{Structures}{\describe{\item{\code{GtkAccessible}}{\emph{undocumented
}}}}
\references{\url{http://developer.gnome.org/doc/API/2.0/gtk/GtkAccessible.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
