\alias{GtkAccessible}
\name{GtkAccessible}
\title{GtkAccessible}
\description{Accessibility support for widgets}
\section{Methods and Functions}{
\code{\link{gtkAccessibleConnectWidgetDestroyed}(object)}\cr

}
\section{Hierarchy}{\preformatted{GObject
   +----AtkObject
         +----GtkAccessible}}
\section{Structures}{\describe{\item{\verb{GtkAccessible}}{
\emph{undocumented
}

}}}
\references{\url{https://developer-old.gnome.org/gtk2/stable/GtkAccessible.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
