\alias{gtkAccessibleConnectWidgetDestroyed}
\name{gtkAccessibleConnectWidgetDestroyed}
\title{gtkAccessibleConnectWidgetDestroyed}
\description{This function specifies the callback function to be called when the widget
corresponding to a GtkAccessible is destroyed.}
\usage{gtkAccessibleConnectWidgetDestroyed(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAccessible}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
