\alias{gtkActionCreateIcon}
\name{gtkActionCreateIcon}
\title{gtkActionCreateIcon}
\description{This function is intended for use by action implementations to
create icons displayed in the proxy widgets.}
\usage{gtkActionCreateIcon(object, icon.size)}
\arguments{
\item{\verb{object}}{the action object}
\item{\verb{icon.size}}{the size of the icon that should be created. \emph{[  \acronym{type}   int]}}
}
\details{Since 2.4}
\value{[\code{\link{GtkWidget}}] a widget that displays the icon for this action.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
