\alias{gtkActionCreateMenu}
\name{gtkActionCreateMenu}
\title{gtkActionCreateMenu}
\description{If \code{action} provides a \code{\link{GtkMenu}} widget as a submenu for the menu
item or the toolbar item it creates, this function returns an
instance of that menu.}
\usage{gtkActionCreateMenu(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAction}}}}
\details{Since 2.12}
\value{[\code{\link{GtkWidget}}] the menu item provided by the action, or \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
