\alias{gtkActionCreateToolItem}
\name{gtkActionCreateToolItem}
\title{gtkActionCreateToolItem}
\description{Creates a toolbar item widget that proxies for the given action.}
\usage{gtkActionCreateToolItem(object)}
\arguments{\item{\verb{object}}{the action object}}
\details{Since 2.4}
\value{[\code{\link{GtkWidget}}] a toolbar item connected to the action.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
