\alias{gtkActionGetAccelClosure}
\name{gtkActionGetAccelClosure}
\title{gtkActionGetAccelClosure}
\description{Returns the accel closure for this action.}
\usage{gtkActionGetAccelClosure(object)}
\arguments{\item{\verb{object}}{the action object}}
\details{Since 2.8}
\value{[\code{\link{GClosure}}] the accel closure for this action.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
