\alias{gtkActionGetName}
\name{gtkActionGetName}
\title{gtkActionGetName}
\description{Returns the name of the action.}
\usage{gtkActionGetName(object)}
\arguments{\item{\verb{object}}{the action object}}
\details{Since 2.4}
\value{[character] the name of the action. The string belongs to GTK+ and should not
be freed.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
