\alias{gtkActionGetProxies}
\name{gtkActionGetProxies}
\title{gtkActionGetProxies}
\description{Returns the proxy widgets for an action.
See also \code{\link{gtkWidgetGetAction}}.}
\usage{gtkActionGetProxies(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkAction}}]  the action object}}
\details{  Since  2.4}
\value{[list]  a \code{list} of proxy widgets.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
