\alias{gtkActionGetProxies}
\name{gtkActionGetProxies}
\title{gtkActionGetProxies}
\description{Returns the proxy widgets for an action.
See also \code{\link{gtkWidgetGetAction}}.}
\usage{gtkActionGetProxies(object)}
\arguments{\item{\verb{object}}{the action object}}
\details{Since 2.4}
\value{[list] a \verb{list} of proxy widgets. \emph{[  \acronym{element-type}   GtkWidget][  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
