\alias{gtkActionGroupListActions}
\name{gtkActionGroupListActions}
\title{gtkActionGroupListActions}
\description{Lists the actions in the action group.}
\usage{gtkActionGroupListActions(object)}
\arguments{\item{\verb{object}}{the action group}}
\details{Since 2.4}
\value{[list] an allocated list of the action objects in the action group. \emph{[  \acronym{element-type}   GtkAction][  \acronym{transfer container}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
