\alias{gtkActionGroupRemoveAction}
\name{gtkActionGroupRemoveAction}
\title{gtkActionGroupRemoveAction}
\description{Removes an action object from the action group.}
\usage{gtkActionGroupRemoveAction(object, action)}
\arguments{
\item{\code{object}}{[\code{\link{GtkActionGroup}}]  the action group}
\item{\code{action}}{[\code{\link{GtkAction}}]  an action}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
