\alias{gtkActionGroupRemoveAction}
\name{gtkActionGroupRemoveAction}
\title{gtkActionGroupRemoveAction}
\description{Removes an action object from the action group.}
\usage{gtkActionGroupRemoveAction(object, action)}
\arguments{
\item{\verb{object}}{the action group}
\item{\verb{action}}{an action}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
