\alias{gtkActionGroupSetSensitive}
\name{gtkActionGroupSetSensitive}
\title{gtkActionGroupSetSensitive}
\description{Changes the sensitivity of \code{action.group}}
\usage{gtkActionGroupSetSensitive(object, sensitive)}
\arguments{
\item{\verb{object}}{the action group}
\item{\verb{sensitive}}{new sensitivity}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
