\alias{gtkActionGroupSetVisible}
\name{gtkActionGroupSetVisible}
\title{gtkActionGroupSetVisible}
\description{Changes the visible of \code{action.group}.}
\usage{gtkActionGroupSetVisible(object, visible)}
\arguments{
\item{\code{object}}{[\code{\link{GtkActionGroup}}]  the action group}
\item{\code{visible}}{[logical]  new visiblity}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
