\alias{gtkActionGroupSetVisible}
\name{gtkActionGroupSetVisible}
\title{gtkActionGroupSetVisible}
\description{Changes the visible of \code{action.group}.}
\usage{gtkActionGroupSetVisible(object, visible)}
\arguments{
\item{\verb{object}}{the action group}
\item{\verb{visible}}{new visiblity}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
