\alias{gtkActionGroupTranslateString}
\name{gtkActionGroupTranslateString}
\title{gtkActionGroupTranslateString}
\description{Translates a string using the specified \code{translateFunc()}. This
is mainly intended for language bindings.}
\usage{gtkActionGroupTranslateString(object, string)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkActionGroup}}}
\item{\verb{string}}{a string}
}
\details{Since 2.6}
\value{[character] the translation of \code{string}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
