\alias{gtkActionIsSensitive}
\name{gtkActionIsSensitive}
\title{gtkActionIsSensitive}
\description{Returns whether the action is effectively sensitive.}
\usage{gtkActionIsSensitive(object)}
\arguments{\item{\verb{object}}{the action object}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if the action and its associated action group
are both sensitive.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
