\alias{gtkActionSetAccelGroup}
\name{gtkActionSetAccelGroup}
\title{gtkActionSetAccelGroup}
\description{Sets the \code{\link{GtkAccelGroup}} in which the accelerator for this action
will be installed.}
\usage{gtkActionSetAccelGroup(object, accel.group)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAction}}]  the action object}
\item{\code{accel.group}}{[\code{\link{GtkAccelGroup}}]  a \code{\link{GtkAccelGroup}} or \code{NULL}}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
