\alias{gtkActionSetAccelGroup}
\name{gtkActionSetAccelGroup}
\title{gtkActionSetAccelGroup}
\description{Sets the \code{\link{GtkAccelGroup}} in which the accelerator for this action
will be installed.}
\usage{gtkActionSetAccelGroup(object, accel.group)}
\arguments{
\item{\verb{object}}{the action object}
\item{\verb{accel.group}}{a \code{\link{GtkAccelGroup}} or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
