\alias{gtkActionSetAccelPath}
\name{gtkActionSetAccelPath}
\title{gtkActionSetAccelPath}
\description{Sets the accel path for this action.  All proxy widgets associated
with the action will have this accel path, so that their
accelerators are consistent.}
\usage{gtkActionSetAccelPath(object, accel.path)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkAction}}]  the action object}
\item{\verb{accel.path}}{[character]  the accelerator path}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
