\alias{gtkActionSetSensitive}
\name{gtkActionSetSensitive}
\title{gtkActionSetSensitive}
\description{Sets the ::sensitive property of the action to \code{sensitive}. Note that 
this doesn't necessarily mean effective sensitivity. See 
\code{\link{gtkActionIsSensitive}} 
for that.}
\usage{gtkActionSetSensitive(object, sensitive)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAction}}]  the action object}
\item{\code{sensitive}}{[logical]  \code{TRUE} to make the action sensitive}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
