\alias{gtkActionSetSensitive}
\name{gtkActionSetSensitive}
\title{gtkActionSetSensitive}
\description{Sets the ::sensitive property of the action to \code{sensitive}. Note that 
this doesn't necessarily mean effective sensitivity. See 
\code{\link{gtkActionIsSensitive}} 
for that.}
\usage{gtkActionSetSensitive(object, sensitive)}
\arguments{
\item{\verb{object}}{the action object}
\item{\verb{sensitive}}{\code{TRUE} to make the action sensitive}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
