\alias{gtkActionSetVisible}
\name{gtkActionSetVisible}
\title{gtkActionSetVisible}
\description{Sets the ::visible property of the action to \code{visible}. Note that 
this doesn't necessarily mean effective visibility. See 
\code{\link{gtkActionIsVisible}} 
for that.}
\usage{gtkActionSetVisible(object, visible)}
\arguments{
\item{\verb{object}}{the action object}
\item{\verb{visible}}{\code{TRUE} to make the action visible}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
