\alias{gtkAdjustmentChanged}
\name{gtkAdjustmentChanged}
\title{gtkAdjustmentChanged}
\description{Emits a "changed" signal from the \verb{\link{GtkAdjustment}}.
This is typically called by the owner of the \verb{\link{GtkAdjustment}} after it has
changed any of the \verb{\link{GtkAdjustment}} fields other than the value.}
\usage{gtkAdjustmentChanged(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkAdjustment}}] a \verb{\link{GtkAdjustment}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
