\alias{gtkAdjustmentChanged}
\name{gtkAdjustmentChanged}
\title{gtkAdjustmentChanged}
\description{Emits a "changed" signal from the \code{\link{GtkAdjustment}}.
This is typically called by the owner of the \code{\link{GtkAdjustment}} after it has
changed any of the \code{\link{GtkAdjustment}} fields other than the value.}
\usage{gtkAdjustmentChanged(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkAdjustment}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
