\alias{gtkAdjustmentNew}
\name{gtkAdjustmentNew}
\title{gtkAdjustmentNew}
\description{Creates a new \code{\link{GtkAdjustment}}.}
\usage{gtkAdjustmentNew(value = NULL, lower = NULL, upper = NULL, step.incr = NULL, page.incr = NULL, page.size = NULL)}
\arguments{
\item{\code{value}}{[numeric] the initial value.}
\item{\code{lower}}{[numeric] the minimum value.}
\item{\code{upper}}{[numeric] the maximum value.}
\item{\code{step.incr}}{[numeric] the step increment.}
\item{\code{page.incr}}{[numeric] the page increment.}
\item{\code{page.size}}{[numeric] the page size.}
}
\value{[\code{\link{GtkObject}}] a new \code{\link{GtkAdjustment}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
