\alias{gtkAdjustmentNew}
\name{gtkAdjustmentNew}
\title{gtkAdjustmentNew}
\description{Creates a new \code{\link{GtkAdjustment}}.}
\usage{gtkAdjustmentNew(value = NULL, lower = NULL, upper = NULL, step.incr = NULL, 
    page.incr = NULL, page.size = NULL)}
\arguments{
\item{\verb{value}}{the initial value.}
\item{\verb{lower}}{the minimum value.}
\item{\verb{upper}}{the maximum value.}
\item{\verb{step.incr}}{the step increment.}
\item{\verb{page.incr}}{the page increment.}
\item{\verb{page.size}}{the page size.}
}
\value{[\code{\link{GtkObject}}] a new \code{\link{GtkAdjustment}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
