\alias{gtkArrowNew}
\name{gtkArrowNew}
\title{gtkArrowNew}
\description{Creates a new arrow widget.}
\usage{gtkArrowNew(arrow.type = NULL, shadow.type = NULL, show = TRUE)}
\arguments{
\item{\code{arrow.type}}{[\code{\link{GtkArrowType}}] a valid \code{\link{GtkArrowType}}.}
\item{\code{shadow.type}}{[\code{\link{GtkShadowType}}] a valid \code{\link{GtkShadowType}}.}
}
\value{[\code{\link{GtkWidget}}] the new \code{\link{GtkArrow}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
