\alias{gtkArrowNew}
\name{gtkArrowNew}
\title{gtkArrowNew}
\description{Creates a new arrow widget.}
\usage{gtkArrowNew(arrow.type = NULL, shadow.type = NULL, show = TRUE)}
\arguments{
\item{\verb{arrow.type}}{a valid \code{\link{GtkArrowType}}.}
\item{\verb{shadow.type}}{a valid \code{\link{GtkShadowType}}.}
}
\value{[\code{\link{GtkWidget}}] the new \code{\link{GtkArrow}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
