\alias{gtkArrowSet}
\name{gtkArrowSet}
\title{gtkArrowSet}
\description{Sets the direction and style of the \code{\link{GtkArrow}}, \code{arrow}.}
\usage{gtkArrowSet(object, arrow.type, shadow.type)}
\arguments{
\item{\code{object}}{[\code{\link{GtkArrow}}] a widget of type \code{\link{GtkArrow}}.}
\item{\code{arrow.type}}{[\code{\link{GtkArrowType}}] a valid \code{\link{GtkArrowType}}.}
\item{\code{shadow.type}}{[\code{\link{GtkShadowType}}] a valid \code{\link{GtkShadowType}}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
