\alias{gtkArrowSet}
\name{gtkArrowSet}
\title{gtkArrowSet}
\description{Sets the direction and style of the \code{\link{GtkArrow}}, \code{arrow}.}
\usage{gtkArrowSet(object, arrow.type, shadow.type)}
\arguments{
\item{\verb{object}}{a widget of type \code{\link{GtkArrow}}.}
\item{\verb{arrow.type}}{a valid \code{\link{GtkArrowType}}.}
\item{\verb{shadow.type}}{a valid \code{\link{GtkShadowType}}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
