\alias{gtkAssistantAddActionWidget}
\name{gtkAssistantAddActionWidget}
\title{gtkAssistantAddActionWidget}
\description{Adds a widget to the action area of a \code{\link{GtkAssistant}}.}
\usage{gtkAssistantAddActionWidget(object, child)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAssistant}}]  a \code{\link{GtkAssistant}}}
\item{\code{child}}{[\code{\link{GtkWidget}}]  a \code{\link{GtkWidget}}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
