\alias{gtkAssistantAddActionWidget}
\name{gtkAssistantAddActionWidget}
\title{gtkAssistantAddActionWidget}
\description{Adds a widget to the action area of a \code{\link{GtkAssistant}}.}
\usage{gtkAssistantAddActionWidget(object, child)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAssistant}}}
\item{\verb{child}}{a \code{\link{GtkWidget}}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
