\alias{gtkAssistantAppendPage}
\name{gtkAssistantAppendPage}
\title{gtkAssistantAppendPage}
\description{Appends a page to the \code{assistant}.}
\usage{gtkAssistantAppendPage(object, page)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAssistant}}}
\item{\verb{page}}{a \code{\link{GtkWidget}}}
}
\details{Since 2.10}
\value{[integer] the index (starting at 0) of the inserted page}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
