\alias{gtkAssistantGetPageHeaderImage}
\name{gtkAssistantGetPageHeaderImage}
\title{gtkAssistantGetPageHeaderImage}
\description{Gets the header image for \code{page}.}
\usage{gtkAssistantGetPageHeaderImage(object, page)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAssistant}}}
\item{\verb{page}}{a page of \code{assistant}}
}
\details{Since 2.10}
\value{[\code{\link{GdkPixbuf}}] the header image for \code{page}, or \code{NULL}
if there's no header image for the page.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
