\alias{gtkAssistantGetPageTitle}
\name{gtkAssistantGetPageTitle}
\title{gtkAssistantGetPageTitle}
\description{Gets the title for \code{page}.}
\usage{gtkAssistantGetPageTitle(object, page)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAssistant}}]  a \code{\link{GtkAssistant}}}
\item{\code{page}}{[\code{\link{GtkWidget}}]  a page of \code{assistant}}
}
\details{  Since  2.10}
\value{[character]  the title for \code{page}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
