\alias{gtkAssistantGetPageType}
\name{gtkAssistantGetPageType}
\title{gtkAssistantGetPageType}
\description{Gets the page type of \code{page}.}
\usage{gtkAssistantGetPageType(object, page)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAssistant}}]  a \code{\link{GtkAssistant}}}
\item{\code{page}}{[\code{\link{GtkWidget}}]  a page of \code{assistant}}
}
\details{  Since  2.10}
\value{[\code{\link{GtkAssistantPageType}}]  the page type of \code{page}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
