\alias{gtkAssistantGetPageType}
\name{gtkAssistantGetPageType}
\title{gtkAssistantGetPageType}
\description{Gets the page type of \code{page}.}
\usage{gtkAssistantGetPageType(object, page)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAssistant}}}
\item{\verb{page}}{a page of \code{assistant}}
}
\details{Since 2.10}
\value{[\code{\link{GtkAssistantPageType}}] the page type of \code{page}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
