\alias{gtkAssistantNew}
\name{gtkAssistantNew}
\title{gtkAssistantNew}
\description{Creates a new \code{\link{GtkAssistant}}.}
\usage{gtkAssistantNew(show = TRUE)}
\details{Since 2.10}
\value{[\code{\link{GtkWidget}}] a newly created \code{\link{GtkAssistant}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
