\alias{gtkAssistantRemoveActionWidget}
\name{gtkAssistantRemoveActionWidget}
\title{gtkAssistantRemoveActionWidget}
\description{Removes a widget from the action area of a \code{\link{GtkAssistant}}.}
\usage{gtkAssistantRemoveActionWidget(object, child)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAssistant}}]  a \code{\link{GtkAssistant}}}
\item{\code{child}}{[\code{\link{GtkWidget}}]  a \code{\link{GtkWidget}}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
