\alias{gtkAssistantRemoveActionWidget}
\name{gtkAssistantRemoveActionWidget}
\title{gtkAssistantRemoveActionWidget}
\description{Removes a widget from the action area of a \code{\link{GtkAssistant}}.}
\usage{gtkAssistantRemoveActionWidget(object, child)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAssistant}}}
\item{\verb{child}}{a \code{\link{GtkWidget}}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
