\alias{gtkAssistantSetPageComplete}
\name{gtkAssistantSetPageComplete}
\title{gtkAssistantSetPageComplete}
\description{Sets whether \code{page} contents are complete. This will make
\code{assistant} update the buttons state to be able to continue the task.}
\usage{gtkAssistantSetPageComplete(object, page, complete)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAssistant}}}
\item{\verb{page}}{a page of \code{assistant}}
\item{\verb{complete}}{the completeness status of the page}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
