\alias{gtkAssistantSetPageHeaderImage}
\name{gtkAssistantSetPageHeaderImage}
\title{gtkAssistantSetPageHeaderImage}
\description{Sets a header image for \code{page}. This image is displayed in the header
area of the assistant when \code{page} is the current page.}
\usage{gtkAssistantSetPageHeaderImage(object, page, pixbuf = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkAssistant}}]  a \code{\link{GtkAssistant}}}
\item{\code{page}}{[\code{\link{GtkWidget}}]  a page of \code{assistant}}
\item{\code{pixbuf}}{[\code{\link{GdkPixbuf}}]  the new header image \code{page}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
