\alias{gtkAssistantSetPageHeaderImage}
\name{gtkAssistantSetPageHeaderImage}
\title{gtkAssistantSetPageHeaderImage}
\description{Sets a header image for \code{page}. This image is displayed in the header
area of the assistant when \code{page} is the current page.}
\usage{gtkAssistantSetPageHeaderImage(object, page, pixbuf = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAssistant}}}
\item{\verb{page}}{a page of \code{assistant}}
\item{\verb{pixbuf}}{the new header image \code{page}. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
