\alias{gtkAssistantSetPageTitle}
\name{gtkAssistantSetPageTitle}
\title{gtkAssistantSetPageTitle}
\description{Sets a title for \code{page}. The title is displayed in the header
area of the assistant when \code{page} is the current page.}
\usage{gtkAssistantSetPageTitle(object, page, title)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkAssistant}}}
\item{\verb{page}}{a page of \code{assistant}}
\item{\verb{title}}{the new title for \code{page}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
