\alias{gtkAssistantSetPageType}
\name{gtkAssistantSetPageType}
\title{gtkAssistantSetPageType}
\description{Sets the page type for \code{page}. The page type determines the page
behavior in the \code{assistant}.}
\usage{gtkAssistantSetPageType(object, page, type)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkAssistant}}]  a \verb{\link{GtkAssistant}}}
\item{\verb{page}}{[\verb{\link{GtkWidget}}]  a page of \code{assistant}}
\item{\verb{type}}{[\code{\link{GtkAssistantPageType}}]  the new type for \code{page}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
