\alias{gtkBorderCopy}
\name{gtkBorderCopy}
\title{gtkBorderCopy}
\description{Copies a \verb{\link{GtkBorder}} structure.}
\usage{gtkBorderCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkBorder}}]  a \verb{\link{GtkBorder}}.}}
\value{[\verb{\link{GtkBorder}}]  a copy of \code{border.}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
