\alias{gtkBorderCopy}
\name{gtkBorderCopy}
\title{gtkBorderCopy}
\description{Copies a \code{\link{GtkBorder}} structure.}
\usage{gtkBorderCopy(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkBorder}}.}}
\value{[\code{\link{GtkBorder}}] a copy of \code{border.}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
