\alias{gtkBoxGetSpacing}
\name{gtkBoxGetSpacing}
\title{gtkBoxGetSpacing}
\description{Gets the value set by \code{\link{gtkBoxSetSpacing}}.}
\usage{gtkBoxGetSpacing(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkBox}}}}
\value{[integer] spacing between children}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
