\alias{gtkBoxSetChildPacking}
\name{gtkBoxSetChildPacking}
\title{gtkBoxSetChildPacking}
\description{Sets the way \code{child} is packed into \code{box}.}
\usage{gtkBoxSetChildPacking(object, child, expand, fill, padding, pack.type)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkBox}}]  a \verb{\link{GtkBox}}}
\item{\verb{child}}{[\verb{\link{GtkWidget}}]  the \verb{\link{GtkWidget}} of the child to set}
\item{\verb{expand}}{[logical]  the new value of the \verb{"expand"} child property }
\item{\verb{fill}}{[logical]  the new value of the \verb{"fill"} child property}
\item{\verb{padding}}{[numeric]  the new value of the \verb{"padding"} child property}
\item{\verb{pack.type}}{[\code{\link{GtkPackType}}]  the new value of the \verb{"pack-type"} child property}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
