\alias{gtkBoxSetChildPacking}
\name{gtkBoxSetChildPacking}
\title{gtkBoxSetChildPacking}
\description{Sets the way \code{child} is packed into \code{box}.}
\usage{gtkBoxSetChildPacking(object, child, expand, fill, padding, pack.type)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkBox}}}
\item{\verb{child}}{the \code{\link{GtkWidget}} of the child to set}
\item{\verb{expand}}{the new value of the \verb{"expand"} child property}
\item{\verb{fill}}{the new value of the \verb{"fill"} child property}
\item{\verb{padding}}{the new value of the \verb{"padding"} child property}
\item{\verb{pack.type}}{the new value of the \verb{"pack-type"} child property}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
