\alias{gtkBoxSetHomogeneous}
\name{gtkBoxSetHomogeneous}
\title{gtkBoxSetHomogeneous}
\description{Sets the \code{:homogeneous} property of \code{box}, controlling 
whether or not all children of \code{box} are given equal space 
in the box.}
\usage{gtkBoxSetHomogeneous(object, homogeneous)}
\arguments{
\item{\code{object}}{[\code{\link{GtkBox}}]  a \code{\link{GtkBox}}}
\item{\code{homogeneous}}{[logical]  a boolean value, \code{TRUE} to create equal allotments,
  \code{FALSE} for variable allotments}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
