\alias{gtkBoxSetSpacing}
\name{gtkBoxSetSpacing}
\title{gtkBoxSetSpacing}
\description{Sets the \verb{"spacing"} property of \code{box}, which is the 
number of pixels to place between children of \code{box}.}
\usage{gtkBoxSetSpacing(object, spacing)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkBox}}}
\item{\verb{spacing}}{the number of pixels to put between children}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
