\alias{gtkBuildableAddChild}
\name{gtkBuildableAddChild}
\title{gtkBuildableAddChild}
\description{Adds a child to \code{buildable}. \code{type} is an optional string
describing how the child should be added.}
\usage{gtkBuildableAddChild(object, builder, child, type)}
\arguments{
\item{\code{object}}{[\code{\link{GtkBuildable}}]  a \code{\link{GtkBuildable}}}
\item{\code{builder}}{[\code{\link{GtkBuilder}}]  a \code{\link{GtkBuilder}}}
\item{\code{child}}{[\code{\link{GObject}}]  child to add}
\item{\code{type}}{[character]  kind of child or \code{NULL}}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
