\alias{gtkBuildableAddChild}
\name{gtkBuildableAddChild}
\title{gtkBuildableAddChild}
\description{Adds a child to \code{buildable}. \code{type} is an optional string
describing how the child should be added.}
\usage{gtkBuildableAddChild(object, builder, child, type)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkBuildable}}}
\item{\verb{builder}}{a \code{\link{GtkBuilder}}}
\item{\verb{child}}{child to add}
\item{\verb{type}}{kind of child or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
