\alias{gtkBuildableGetInternalChild}
\name{gtkBuildableGetInternalChild}
\title{gtkBuildableGetInternalChild}
\description{Get the internal child called \code{childname} of the \code{buildable} object.}
\usage{gtkBuildableGetInternalChild(object, builder, childname)}
\arguments{
\item{\code{object}}{[\code{\link{GtkBuildable}}]  a \code{\link{GtkBuildable}}}
\item{\code{builder}}{[\code{\link{GtkBuilder}}]  a \code{\link{GtkBuilder}}}
\item{\code{childname}}{[character]  name of child}
}
\details{  Since  2.12}
\value{[\code{\link{GObject}}]  the internal child of the buildable object }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
