\alias{gtkBuildableGetInternalChild}
\name{gtkBuildableGetInternalChild}
\title{gtkBuildableGetInternalChild}
\description{Get the internal child called \code{childname} of the \code{buildable} object.}
\usage{gtkBuildableGetInternalChild(object, builder, childname)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkBuildable}}}
\item{\verb{builder}}{a \code{\link{GtkBuilder}}}
\item{\verb{childname}}{name of child}
}
\details{Since 2.12}
\value{[\code{\link{GObject}}] the internal child of the buildable object}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
