\alias{gtkBuildableGetName}
\name{gtkBuildableGetName}
\title{gtkBuildableGetName}
\description{Gets the name of the \code{buildable} object. }
\usage{gtkBuildableGetName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkBuildable}}}}
\details{\code{\link{GtkBuilder}} sets the name based on the the 
GtkBuilder UI definition 
used to construct the \code{buildable}.
  Since 2.12}
\value{[character] the name set with \code{\link{gtkBuildableSetName}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
