\alias{gtkBuildableSetBuildableProperty}
\name{gtkBuildableSetBuildableProperty}
\title{gtkBuildableSetBuildableProperty}
\description{Sets the property name \code{name} to \code{value} on the \code{buildable} object.}
\usage{gtkBuildableSetBuildableProperty(object, builder, name, value)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkBuildable}}}
\item{\verb{builder}}{a \code{\link{GtkBuilder}}}
\item{\verb{name}}{name of property}
\item{\verb{value}}{value of property}
}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
