\alias{gtkBuildableSetName}
\name{gtkBuildableSetName}
\title{gtkBuildableSetName}
\description{Sets the name of the \code{buildable} object.}
\usage{gtkBuildableSetName(object, name)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkBuildable}}]  a \verb{\link{GtkBuildable}}}
\item{\verb{name}}{[character]  name to set}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
