\alias{gtkBuildableSetName}
\name{gtkBuildableSetName}
\title{gtkBuildableSetName}
\description{Sets the name of the \code{buildable} object.}
\usage{gtkBuildableSetName(object, name)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkBuildable}}}
\item{\verb{name}}{name to set}
}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
