\alias{gtkBuilderGetTranslationDomain}
\name{gtkBuilderGetTranslationDomain}
\title{gtkBuilderGetTranslationDomain}
\description{Gets the translation domain of \code{builder}.}
\usage{gtkBuilderGetTranslationDomain(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkBuilder}}]  a \verb{\link{GtkBuilder}}}}
\details{  Since  2.12}
\value{[character]  the translation domain.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
