\alias{gtkBuilderGetTranslationDomain}
\name{gtkBuilderGetTranslationDomain}
\title{gtkBuilderGetTranslationDomain}
\description{Gets the translation domain of \code{builder}.}
\usage{gtkBuilderGetTranslationDomain(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkBuilder}}}}
\details{Since 2.12}
\value{[character] the translation domain. This string is owned
by the builder object and must not be modified or freed.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
