\alias{gtkBuilderNew}
\name{gtkBuilderNew}
\title{gtkBuilderNew}
\description{Creates a new builder object.}
\usage{gtkBuilderNew()}
\details{  Since  2.12}
\value{[\verb{\link{GtkBuilder}}]  a new \verb{\link{GtkBuilder}} object}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
