\alias{gtkBuilderSetTranslationDomain}
\name{gtkBuilderSetTranslationDomain}
\title{gtkBuilderSetTranslationDomain}
\description{Sets the translation domain of \code{builder}. 
See \verb{"translation-domain"}.}
\usage{gtkBuilderSetTranslationDomain(object, domain)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkBuilder}}}
\item{\verb{domain}}{the translation domain or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
