\alias{gtkButtonBoxSetLayout}
\name{gtkButtonBoxSetLayout}
\title{gtkButtonBoxSetLayout}
\description{Changes the way buttons are arranged in their container.}
\usage{gtkButtonBoxSetLayout(object, layout.style)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkButtonBox}}.}
\item{\verb{layout.style}}{the new layout style.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
