\alias{gtkButtonClicked}
\name{gtkButtonClicked}
\title{gtkButtonClicked}
\description{Emits a \code{\link{gtkButtonClicked}} signal to the given \code{\link{GtkButton}}.}
\usage{gtkButtonClicked(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkButton}}] The \code{\link{GtkButton}} you want to send the signal to.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
