\alias{gtkButtonEnter}
\name{gtkButtonEnter}
\title{gtkButtonEnter}
\description{Emits a \code{\link{GtkButton}}::enter signal to the given \code{\link{GtkButton}}.}
\usage{gtkButtonEnter(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkButton}}] The \code{\link{GtkButton}} you want to send the signal to.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
