\alias{gtkButtonEnter}
\name{gtkButtonEnter}
\title{gtkButtonEnter}
\description{
Emits a \code{\link{gtkButtonEnter}} signal to the given \code{\link{GtkButton}}.
\strong{WARNING: \code{gtk_button_enter} has been deprecated since version 2.20 and should not be used in newly-written code. Use the \verb{"enter-notify-event"} signal.}
}
\usage{gtkButtonEnter(object)}
\arguments{\item{\verb{object}}{The \code{\link{GtkButton}} you want to send the signal to.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
