\alias{gtkButtonGetAlignment}
\name{gtkButtonGetAlignment}
\title{gtkButtonGetAlignment}
\description{Gets the alignment of the child in the button.}
\usage{gtkButtonGetAlignment(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkButton}}}}
\details{Since 2.4}
\value{
A list containing the following elements:
\item{\verb{xalign}}{return location for horizontal alignment}
\item{\verb{yalign}}{return location for vertical alignment}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
