\alias{gtkButtonGetRelief}
\name{gtkButtonGetRelief}
\title{gtkButtonGetRelief}
\description{Returns the current relief style of the given \code{\link{GtkButton}}.}
\usage{gtkButtonGetRelief(object)}
\arguments{\item{\verb{object}}{The \code{\link{GtkButton}} you want the \code{\link{GtkReliefStyle}} from.}}
\value{[\code{\link{GtkReliefStyle}}] The current \code{\link{GtkReliefStyle}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
