\alias{gtkButtonLeave}
\name{gtkButtonLeave}
\title{gtkButtonLeave}
\description{Emits a \code{\link{GtkButton}}::leave signal to the given \code{\link{GtkButton}}.}
\usage{gtkButtonLeave(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkButton}}] The \code{\link{GtkButton}} you want to send the signal to.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
