\alias{gtkButtonLeave}
\name{gtkButtonLeave}
\title{gtkButtonLeave}
\description{
Emits a \code{\link{gtkButtonLeave}} signal to the given \code{\link{GtkButton}}.
\strong{WARNING: \code{gtk_button_leave} has been deprecated since version 2.20 and should not be used in newly-written code. Use the \verb{"leave-notify-event"} signal.}
}
\usage{gtkButtonLeave(object)}
\arguments{\item{\verb{object}}{The \code{\link{GtkButton}} you want to send the signal to.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
