\alias{gtkButtonNew}
\name{gtkButtonNew}
\title{gtkButtonNew}
\description{Creates a new \code{\link{GtkButton}} widget. To add a child widget to the button,
use \code{\link{gtkContainerAdd}}.}
\usage{gtkButtonNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] The newly created \code{\link{GtkButton}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
