\alias{gtkButtonNewWithLabel}
\name{gtkButtonNewWithLabel}
\title{gtkButtonNewWithLabel}
\description{Creates a \code{\link{GtkButton}} widget with a \code{\link{GtkLabel}} child containing the given
text.}
\usage{gtkButtonNewWithLabel(label, show = TRUE)}
\arguments{\item{\verb{label}}{The text you want the \code{\link{GtkLabel}} to hold.}}
\value{[\code{\link{GtkWidget}}] The newly created \code{\link{GtkButton}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
