\alias{gtkButtonPressed}
\name{gtkButtonPressed}
\title{gtkButtonPressed}
\description{Emits a \code{\link{gtkButtonPressed}} signal to the given \verb{\link{GtkButton}}.}
\usage{gtkButtonPressed(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkButton}}] The \verb{\link{GtkButton}} you want to send the signal to.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
