\alias{gtkButtonPressed}
\name{gtkButtonPressed}
\title{gtkButtonPressed}
\description{Emits a \code{\link{gtkButtonPressed}} signal to the given \code{\link{GtkButton}}.}
\usage{gtkButtonPressed(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkButton}}] The \code{\link{GtkButton}} you want to send the signal to.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
