\alias{gtkButtonPressed}
\name{gtkButtonPressed}
\title{gtkButtonPressed}
\description{
Emits a \code{\link{gtkButtonPressed}} signal to the given \code{\link{GtkButton}}.
\strong{WARNING: \code{gtk_button_pressed} has been deprecated since version 2.20 and should not be used in newly-written code. Use the \verb{"button-press-event"} signal.}
}
\usage{gtkButtonPressed(object)}
\arguments{\item{\verb{object}}{The \code{\link{GtkButton}} you want to send the signal to.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
