\alias{gtkButtonReleased}
\name{gtkButtonReleased}
\title{gtkButtonReleased}
\description{Emits a \code{\link{gtkButtonReleased}} signal to the given \code{\link{GtkButton}}.}
\usage{gtkButtonReleased(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkButton}}] The \code{\link{GtkButton}} you want to send the signal to.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
