\alias{gtkButtonReleased}
\name{gtkButtonReleased}
\title{gtkButtonReleased}
\description{
Emits a \code{\link{gtkButtonReleased}} signal to the given \code{\link{GtkButton}}.
\strong{WARNING: \code{gtk_button_released} has been deprecated since version 2.20 and should not be used in newly-written code. Use the \verb{"button-release-event"} signal.}
}
\usage{gtkButtonReleased(object)}
\arguments{\item{\verb{object}}{The \code{\link{GtkButton}} you want to send the signal to.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
