\alias{gtkButtonSetAlignment}
\name{gtkButtonSetAlignment}
\title{gtkButtonSetAlignment}
\description{Sets the alignment of the child. This property has no effect unless 
the child is a \code{\link{GtkMisc}} or a \verb{GtkAligment}.}
\usage{gtkButtonSetAlignment(object, xalign, yalign)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkButton}}}
\item{\verb{xalign}}{the horizontal position of the child, 0.0 is left aligned,
1.0 is right aligned}
\item{\verb{yalign}}{the vertical position of the child, 0.0 is top aligned,
1.0 is bottom aligned}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
