\alias{gtkButtonSetImagePosition}
\name{gtkButtonSetImagePosition}
\title{gtkButtonSetImagePosition}
\description{Sets the position of the image relative to the text 
inside the button.}
\usage{gtkButtonSetImagePosition(object, position)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkButton}}}
\item{\verb{position}}{the position}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
