\alias{gtkButtonSetRelief}
\name{gtkButtonSetRelief}
\title{gtkButtonSetRelief}
\description{Sets the relief style of the edges of the given \code{\link{GtkButton}} widget.
Three styles exist, GTK_RELIEF_NORMAL, GTK_RELIEF_HALF, GTK_RELIEF_NONE.
The default style is, as one can guess, GTK_RELIEF_NORMAL.}
\usage{gtkButtonSetRelief(object, newstyle)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkButton}} you want to set relief styles of.}
\item{\verb{newstyle}}{The GtkReliefStyle as described above.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
