\alias{gtkButtonSetUseStock}
\name{gtkButtonSetUseStock}
\title{gtkButtonSetUseStock}
\description{If \code{TRUE}, the label set on the button is used as a
stock id to select the stock item for the button.}
\usage{gtkButtonSetUseStock(object, use.stock)}
\arguments{
\item{\code{object}}{[\code{\link{GtkButton}}]  a \code{\link{GtkButton}}}
\item{\code{use.stock}}{[logical]  \code{TRUE} if the button should use a stock item}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
