\alias{gtkCalendarClearMarks}
\name{gtkCalendarClearMarks}
\title{gtkCalendarClearMarks}
\description{Remove all visual markers.}
\usage{gtkCalendarClearMarks(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkCalendar}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
