\alias{gtkCalendarDisplayOptions}
\name{gtkCalendarDisplayOptions}
\title{gtkCalendarDisplayOptions}
\description{
Sets display options (whether to display the heading and the month headings).
\strong{WARNING: \code{gtk_calendar_display_options} is deprecated and should not be used in newly-written code. }
}
\usage{gtkCalendarDisplayOptions(object, flags)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCalendar}}]  a \verb{\link{GtkCalendar}}.}
\item{\verb{flags}}{[\code{\link{GtkCalendarDisplayOptions}}]  the display options to set.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
