\alias{gtkCalendarDisplayOptions}
\name{gtkCalendarDisplayOptions}
\title{gtkCalendarDisplayOptions}
\description{
Sets display options (whether to display the heading and the month headings).
\strong{WARNING: \code{gtk_calendar_display_options} has been deprecated since version 2.4 and should not be used in newly-written code. Use \code{\link{gtkCalendarSetDisplayOptions}} instead}
}
\usage{gtkCalendarDisplayOptions(object, flags)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkCalendar}}.}
\item{\verb{flags}}{the display options to set.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
