\alias{gtkCalendarFreeze}
\name{gtkCalendarFreeze}
\title{gtkCalendarFreeze}
\description{
Does nothing. Previously locked the display of the calendar until
it was thawed with \code{\link{gtkCalendarThaw}}.
\strong{WARNING: \code{gtk_calendar_freeze} has been deprecated since version 2.8 and should not be used in newly-written code. }
}
\usage{gtkCalendarFreeze(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkCalendar}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
