\alias{gtkCalendarMarkDay}
\name{gtkCalendarMarkDay}
\title{gtkCalendarMarkDay}
\description{Places a visual marker on a particular day.}
\usage{gtkCalendarMarkDay(object, day)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCalendar}}]  a \verb{\link{GtkCalendar}} }
\item{\verb{day}}{[numeric]  the day number to mark between 1 and 31.}
}
\value{[logical] \code{TRUE}, always}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
