\alias{gtkCalendarMarkDay}
\name{gtkCalendarMarkDay}
\title{gtkCalendarMarkDay}
\description{Places a visual marker on a particular day.}
\usage{gtkCalendarMarkDay(object, day)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkCalendar}}}
\item{\verb{day}}{the day number to mark between 1 and 31.}
}
\value{[logical] \code{TRUE}, always}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
