\alias{gtkCalendarSelectDay}
\name{gtkCalendarSelectDay}
\title{gtkCalendarSelectDay}
\description{Selects a day from the current month.}
\usage{gtkCalendarSelectDay(object, day)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCalendar}}]  a \verb{\link{GtkCalendar}}.}
\item{\verb{day}}{[numeric]  the day number between 1 and 31, or 0 to unselect 
  the currently selected day.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
