\alias{gtkCalendarSelectMonth}
\name{gtkCalendarSelectMonth}
\title{gtkCalendarSelectMonth}
\description{Shifts the calendar to a different month.}
\usage{gtkCalendarSelectMonth(object, month, year)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkCalendar}}}
\item{\verb{month}}{a month number between 0 and 11.}
\item{\verb{year}}{the year the month is in.}
}
\value{[logical] \code{TRUE}, always}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
