\alias{gtkCalendarSetDisplayOptions}
\name{gtkCalendarSetDisplayOptions}
\title{gtkCalendarSetDisplayOptions}
\description{Sets display options (whether to display the heading and the month  
headings).}
\usage{gtkCalendarSetDisplayOptions(object, flags)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCalendar}}]  a \code{\link{GtkCalendar}}}
\item{\code{flags}}{[\code{\link{GtkCalendarDisplayOptions}}]  the display options to set}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
