\alias{gtkCalendarSetDisplayOptions}
\name{gtkCalendarSetDisplayOptions}
\title{gtkCalendarSetDisplayOptions}
\description{Sets display options (whether to display the heading and the month  
headings).}
\usage{gtkCalendarSetDisplayOptions(object, flags)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkCalendar}}}
\item{\verb{flags}}{the display options to set}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
