\alias{gtkCalendarThaw}
\name{gtkCalendarThaw}
\title{gtkCalendarThaw}
\description{
Does nothing. Previously defrosted a calendar; all the changes made
since the last \code{\link{gtkCalendarFreeze}} were displayed.
\strong{WARNING: \code{gtk_calendar_thaw} has been deprecated since version 2.8 and should not be used in newly-written code. }
}
\usage{gtkCalendarThaw(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkCalendar}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
