\alias{gtkCalendarUnmarkDay}
\name{gtkCalendarUnmarkDay}
\title{gtkCalendarUnmarkDay}
\description{Removes the visual marker from a particular day.}
\usage{gtkCalendarUnmarkDay(object, day)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkCalendar}}.}
\item{\verb{day}}{the day number to unmark between 1 and 31.}
}
\value{[logical] \code{TRUE}, always}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
