\alias{gtkCellEditableEditingDone}
\name{gtkCellEditableEditingDone}
\title{gtkCellEditableEditingDone}
\description{Emits the \code{\link{gtkCellEditableEditingDone}} signal.}
\usage{gtkCellEditableEditingDone(object)}
\arguments{\item{\verb{object}}{A \verb{GtkTreeEditable}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
