\alias{gtkCellEditableRemoveWidget}
\name{gtkCellEditableRemoveWidget}
\title{gtkCellEditableRemoveWidget}
\description{Emits the \code{\link{gtkCellEditableRemoveWidget}} signal.}
\usage{gtkCellEditableRemoveWidget(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkCellEditable}}]  A \verb{GtkTreeEditable}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
