\alias{gtkCellEditableRemoveWidget}
\name{gtkCellEditableRemoveWidget}
\title{gtkCellEditableRemoveWidget}
\description{Emits the "remove\_widget" signal.  This signal is meant to indicate that the
cell is finished editing, and the widget may now be destroyed.}
\usage{gtkCellEditableRemoveWidget(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkCellEditable}}]  A \code{GtkTreeEditable}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
