\alias{gtkCellLayoutClearAttributes}
\name{gtkCellLayoutClearAttributes}
\title{gtkCellLayoutClearAttributes}
\description{Clears all existing attributes previously set with
\code{\link{gtkCellLayoutSetAttributes}}.}
\usage{gtkCellLayoutClearAttributes(object, cell)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkCellLayout}}.}
\item{\verb{cell}}{A \code{\link{GtkCellRenderer}} to clear the attribute mapping on.}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
