\alias{gtkCellLayoutGetCells}
\name{gtkCellLayoutGetCells}
\title{gtkCellLayoutGetCells}
\description{Returns the cell renderers which have been added to \code{cell.layout}.}
\usage{gtkCellLayoutGetCells(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkCellLayout}}}}
\details{Since 2.12}
\value{[list] a list of cell renderers. The list, \emph{[  \acronym{element-type}   GtkCellRenderer][  \acronym{transfer container}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
