\alias{gtkCellLayoutReorder}
\name{gtkCellLayoutReorder}
\title{gtkCellLayoutReorder}
\description{Re-inserts \code{cell} at \code{position}. Note that \code{cell} has already to be packed
into \code{cell.layout} for this to function properly.}
\usage{gtkCellLayoutReorder(object, cell, position)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkCellLayout}}.}
\item{\verb{cell}}{A \code{\link{GtkCellRenderer}} to reorder.}
\item{\verb{position}}{New position to insert \code{cell} at.}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
