\alias{gtkCellRendererAccelNew}
\name{gtkCellRendererAccelNew}
\title{gtkCellRendererAccelNew}
\description{Creates a new \verb{\link{GtkCellRendererAccel}}.}
\usage{gtkCellRendererAccelNew()}
\details{  Since  2.10}
\value{[\verb{\link{GtkCellRenderer}}]  the new cell renderer}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
