\alias{gtkCellRendererGetFixedSize}
\name{gtkCellRendererGetFixedSize}
\title{gtkCellRendererGetFixedSize}
\description{Fills in \code{width} and \code{height} with the appropriate size of \code{cell}.}
\usage{gtkCellRendererGetFixedSize(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkCellRenderer}}}}
\value{
A list containing the following elements:
\item{\verb{width}}{location to fill in with the fixed width of the cell, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
\item{\verb{height}}{location to fill in with the fixed height of the cell, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
