\alias{gtkCellRendererProgressNew}
\name{gtkCellRendererProgressNew}
\title{gtkCellRendererProgressNew}
\description{Creates a new \verb{\link{GtkCellRendererProgress}}.}
\usage{gtkCellRendererProgressNew()}
\details{  Since  2.6}
\value{[\verb{\link{GtkCellRenderer}}]  the new cell renderer}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
