\alias{gtkCellRendererProgressNew}
\name{gtkCellRendererProgressNew}
\title{gtkCellRendererProgressNew}
\description{Creates a new \code{\link{GtkCellRendererProgress}}.}
\usage{gtkCellRendererProgressNew()}
\details{Since 2.6}
\value{[\code{\link{GtkCellRenderer}}] the new cell renderer}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
