\alias{gtkCellRendererSetFixedSize}
\name{gtkCellRendererSetFixedSize}
\title{gtkCellRendererSetFixedSize}
\description{Sets the renderer size to be explicit, independent of the properties set.}
\usage{gtkCellRendererSetFixedSize(object, width, height)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCellRenderer}}]  A \code{\link{GtkCellRenderer}}}
\item{\code{width}}{[integer]  the width of the cell renderer, or -1}
\item{\code{height}}{[integer]  the height of the cell renderer, or -1}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
