\alias{gtkCellRendererSetFixedSize}
\name{gtkCellRendererSetFixedSize}
\title{gtkCellRendererSetFixedSize}
\description{Sets the renderer size to be explicit, independent of the properties set.}
\usage{gtkCellRendererSetFixedSize(object, width, height)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkCellRenderer}}}
\item{\verb{width}}{the width of the cell renderer, or -1}
\item{\verb{height}}{the height of the cell renderer, or -1}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
