\alias{gtkCellRendererSpinNew}
\name{gtkCellRendererSpinNew}
\title{gtkCellRendererSpinNew}
\description{Creates a new \code{\link{GtkCellRendererSpin}}.}
\usage{gtkCellRendererSpinNew()}
\details{Since 2.10}
\value{[\code{\link{GtkCellRenderer}}] a new \code{\link{GtkCellRendererSpin}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
