\alias{gtkCellRendererToggleGetRadio}
\name{gtkCellRendererToggleGetRadio}
\title{gtkCellRendererToggleGetRadio}
\description{Returns whether we're rendering radio toggles rather than checkboxes.}
\usage{gtkCellRendererToggleGetRadio(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkCellRendererToggle}}}}
\value{[logical] \code{TRUE} if we're rendering radio toggles rather than checkboxes}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
