\alias{gtkCellRendererToggleNew}
\name{gtkCellRendererToggleNew}
\title{gtkCellRendererToggleNew}
\description{Creates a new \code{\link{GtkCellRendererToggle}}. Adjust rendering
parameters using object properties. Object properties can be set
globally (with \code{\link{gObjectSet}}). Also, with \code{\link{GtkTreeViewColumn}}, you
can bind a property to a value in a \code{\link{GtkTreeModel}}. For example, you
can bind the "active" property on the cell renderer to a boolean value
in the model, thus causing the check button to reflect the state of
the model.}
\usage{gtkCellRendererToggleNew()}
\value{[\code{\link{GtkCellRenderer}}] the new cell renderer}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
