\alias{gtkCellRendererToggleSetActive}
\name{gtkCellRendererToggleSetActive}
\title{gtkCellRendererToggleSetActive}
\description{Activates or deactivates a cell renderer.}
\usage{gtkCellRendererToggleSetActive(object, setting)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCellRendererToggle}}]  a \code{\link{GtkCellRendererToggle}}.}
\item{\code{setting}}{[logical]  the value to set.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
