\alias{gtkCellRendererToggleSetActive}
\name{gtkCellRendererToggleSetActive}
\title{gtkCellRendererToggleSetActive}
\description{Activates or deactivates a cell renderer.}
\usage{gtkCellRendererToggleSetActive(object, setting)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkCellRendererToggle}}.}
\item{\verb{setting}}{the value to set.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
