\alias{gtkCellViewGetCellRenderers}
\name{gtkCellViewGetCellRenderers}
\title{gtkCellViewGetCellRenderers}
\description{Returns the cell renderers which have been added to \code{cell.view}.}
\usage{gtkCellViewGetCellRenderers(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkCellView}}]  a \verb{\link{GtkCellView}}}}
\details{  Since  2.6}
\value{[list]  a list of cell renderers. The list.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
