\alias{gtkCellViewGetCellRenderers}
\name{gtkCellViewGetCellRenderers}
\title{gtkCellViewGetCellRenderers}
\description{
Returns the cell renderers which have been added to \code{cell.view}.
\strong{WARNING: \code{gtk_cell_view_get_cell_renderers} has been deprecated since version 2.18 and should not be used in newly-written code. use \code{\link{gtkCellLayoutGetCells}} instead.}
}
\usage{gtkCellViewGetCellRenderers(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkCellView}}}}
\details{Since 2.6}
\value{[list] a list of cell renderers. The list,}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
