\alias{gtkCellViewGetDisplayedRow}
\name{gtkCellViewGetDisplayedRow}
\title{gtkCellViewGetDisplayedRow}
\description{Returns a \code{\link{GtkTreePath}} referring to the currently 
displayed row. If no row is currently displayed, 
\code{NULL} is returned.}
\usage{gtkCellViewGetDisplayedRow(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkCellView}}}}
\details{Since 2.6}
\value{[\code{\link{GtkTreePath}}] the currently displayed row or \code{NULL}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
