\alias{gtkCellViewNew}
\name{gtkCellViewNew}
\title{gtkCellViewNew}
\description{Creates a new \verb{\link{GtkCellView}} widget.}
\usage{gtkCellViewNew(show = TRUE)}
\details{  Since  2.6}
\value{[\verb{\link{GtkWidget}}]  A newly created \verb{\link{GtkCellView}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
