\alias{gtkCellViewNewWithMarkup}
\name{gtkCellViewNewWithMarkup}
\title{gtkCellViewNewWithMarkup}
\description{Creates a new \code{\link{GtkCellView}} widget, adds a \code{\link{GtkCellRendererText}} 
to it, and makes it show \code{markup}. The text can be
marked up with the Pango text 
markup language.}
\usage{gtkCellViewNewWithMarkup(markup)}
\arguments{\item{\verb{markup}}{the text to display in the cell view}}
\details{Since 2.6}
\value{[\code{\link{GtkWidget}}] A newly created \code{\link{GtkCellView}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
