\alias{gtkCellViewNewWithPixbuf}
\name{gtkCellViewNewWithPixbuf}
\title{gtkCellViewNewWithPixbuf}
\description{Creates a new \code{\link{GtkCellView}} widget, adds a \code{\link{GtkCellRendererPixbuf}} 
to it, and makes its show \code{pixbuf}.}
\usage{gtkCellViewNewWithPixbuf(pixbuf)}
\arguments{\item{\verb{pixbuf}}{the image to display in the cell view}}
\details{Since 2.6}
\value{[\code{\link{GtkWidget}}] A newly created \code{\link{GtkCellView}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
