\alias{gtkCellViewNewWithText}
\name{gtkCellViewNewWithText}
\title{gtkCellViewNewWithText}
\description{Creates a new \code{\link{GtkCellView}} widget, adds a \code{\link{GtkCellRendererText}} 
to it, and makes its show \code{text}.}
\usage{gtkCellViewNewWithText(text)}
\arguments{\item{\verb{text}}{the text to display in the cell view}}
\details{Since 2.6}
\value{[\code{\link{GtkWidget}}] A newly created \code{\link{GtkCellView}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
