\alias{gtkCellViewSetBackgroundColor}
\name{gtkCellViewSetBackgroundColor}
\title{gtkCellViewSetBackgroundColor}
\description{Sets the background color of \code{view}.}
\usage{gtkCellViewSetBackgroundColor(object, color)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCellView}}]  a \verb{\link{GtkCellView}}}
\item{\verb{color}}{[\verb{\link{GdkColor}}]  the new background color}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
