\alias{gtkCellViewSetBackgroundColor}
\name{gtkCellViewSetBackgroundColor}
\title{gtkCellViewSetBackgroundColor}
\description{Sets the background color of \code{view}.}
\usage{gtkCellViewSetBackgroundColor(object, color)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkCellView}}}
\item{\verb{color}}{the new background color}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
