\alias{gtkCheckButtonNew}
\name{gtkCheckButtonNew}
\title{gtkCheckButtonNew}
\description{Creates a new \verb{\link{GtkCheckButton}}.}
\usage{gtkCheckButtonNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}] a \verb{\link{GtkWidget}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
