\alias{gtkCheckButtonNewWithLabel}
\name{gtkCheckButtonNewWithLabel}
\title{gtkCheckButtonNewWithLabel}
\description{Creates a new \code{\link{GtkCheckButton}} with a \code{\link{GtkLabel}} to the right of it.}
\usage{gtkCheckButtonNewWithLabel(label, show = TRUE)}
\arguments{\item{\verb{label}}{the text for the check button.}}
\value{[\code{\link{GtkWidget}}] a \code{\link{GtkWidget}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
