\alias{gtkCheckMenuItemGetActive}
\name{gtkCheckMenuItemGetActive}
\title{gtkCheckMenuItemGetActive}
\description{Returns whether the check menu item is active. See
\code{\link{gtkCheckMenuItemSetActive}}.}
\usage{gtkCheckMenuItemGetActive(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkCheckMenuItem}}}}
\value{[logical] \code{TRUE} if the menu item is checked.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
