\alias{gtkCheckMenuItemGetDrawAsRadio}
\name{gtkCheckMenuItemGetDrawAsRadio}
\title{gtkCheckMenuItemGetDrawAsRadio}
\description{Returns whether \code{check.menu.item} looks like a \code{\link{GtkRadioMenuItem}}}
\usage{gtkCheckMenuItemGetDrawAsRadio(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkCheckMenuItem}}}}
\details{Since 2.4}
\value{[logical] Whether \code{check.menu.item} looks like a \code{\link{GtkRadioMenuItem}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
