\alias{gtkCheckMenuItemGetInconsistent}
\name{gtkCheckMenuItemGetInconsistent}
\title{gtkCheckMenuItemGetInconsistent}
\description{Retrieves the value set by \code{\link{gtkCheckMenuItemSetInconsistent}}.}
\usage{gtkCheckMenuItemGetInconsistent(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkCheckMenuItem}}}}
\value{[logical] \code{TRUE} if inconsistent}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
