\alias{gtkCheckMenuItemNew}
\name{gtkCheckMenuItemNew}
\title{gtkCheckMenuItemNew}
\description{Creates a new \verb{\link{GtkCheckMenuItem}}.}
\usage{gtkCheckMenuItemNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}] a new \verb{\link{GtkCheckMenuItem}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
