\alias{gtkCheckMenuItemNewWithLabel}
\name{gtkCheckMenuItemNewWithLabel}
\title{gtkCheckMenuItemNewWithLabel}
\description{Creates a new \code{\link{GtkCheckMenuItem}} with a label.}
\usage{gtkCheckMenuItemNewWithLabel(label, show = TRUE)}
\arguments{\item{\verb{label}}{the string to use for the label.}}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkCheckMenuItem}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
