\alias{gtkCheckMenuItemSetDrawAsRadio}
\name{gtkCheckMenuItemSetDrawAsRadio}
\title{gtkCheckMenuItemSetDrawAsRadio}
\description{Sets whether \code{check.menu.item} is drawn like a \code{\link{GtkRadioMenuItem}}}
\usage{gtkCheckMenuItemSetDrawAsRadio(object, draw.as.radio)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkCheckMenuItem}}}
\item{\verb{draw.as.radio}}{whether \code{check.menu.item} is drawn like a \code{\link{GtkRadioMenuItem}}}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
