\alias{gtkCheckMenuItemToggled}
\name{gtkCheckMenuItemToggled}
\title{gtkCheckMenuItemToggled}
\description{Emits the GtkCheckMenuItem::toggled signal.}
\usage{gtkCheckMenuItemToggled(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkCheckMenuItem}}] a \verb{\link{GtkCheckMenuItem}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
