\alias{gtkClipboardGetDisplay}
\name{gtkClipboardGetDisplay}
\title{gtkClipboardGetDisplay}
\description{Gets the \code{\link{GdkDisplay}} associated with \code{clipboard}}
\usage{gtkClipboardGetDisplay(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkClipboard}}}}
\details{Since 2.2}
\value{[\code{\link{GdkDisplay}}] the \code{\link{GdkDisplay}} associated with \code{clipboard}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
