\alias{gtkClipboardStore}
\name{gtkClipboardStore}
\title{gtkClipboardStore}
\description{Stores the current clipboard data somewhere so that it will stay
around after the application has quit.}
\usage{gtkClipboardStore(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkClipboard}}}}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
