\alias{gtkCListClear}
\name{gtkCListClear}
\title{gtkCListClear}
\description{
Removes all the CList's rows.
\strong{WARNING: \code{gtk_clist_clear} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListClear(object)}
\arguments{\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
