\alias{gtkCListGetColumnTitle}
\name{gtkCListGetColumnTitle}
\title{gtkCListGetColumnTitle}
\description{
Gets the current title of the specified column
\strong{WARNING: \code{gtk_clist_get_column_title} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListGetColumnTitle(object, column)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{column}}{The column to query.}
}
\value{[character] The title of the column.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
