\alias{gtkCListGetColumnWidget}
\name{gtkCListGetColumnWidget}
\title{gtkCListGetColumnWidget}
\description{
Gets the widget in the column header for the specified column.
\strong{WARNING: \code{gtk_clist_get_column_widget} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListGetColumnWidget(object, column)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{column}}{The column to query.}
}
\value{[\code{\link{GtkWidget}}] Pointer to a \code{\link{GtkWidget}} for the column header.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
