\alias{gtkCListGetRowStyle}
\name{gtkCListGetRowStyle}
\title{gtkCListGetRowStyle}
\description{
Gets the style set for the specified row.
\strong{WARNING: \code{gtk_clist_get_row_style} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListGetRowStyle(object, row)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{row}}{The row to query.}
}
\value{[\code{\link{GtkStyle}}] The \code{\link{GtkStyle}} of the row.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
