\alias{gtkCListGetSelectable}
\name{gtkCListGetSelectable}
\title{gtkCListGetSelectable}
\description{
Gets whether the specified row is selectable or not.
\strong{WARNING: \code{gtk_clist_get_selectable} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListGetSelectable(object, row)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{row}}{The row to query.}
}
\value{[logical] A \verb{logical} value.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
