\alias{gtkCListGetVadjustment}
\name{gtkCListGetVadjustment}
\title{gtkCListGetVadjustment}
\description{
Gets the \code{\link{GtkAdjustment}} currently being used for the vertical
aspect.
\strong{WARNING: \code{gtk_clist_get_vadjustment} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListGetVadjustment(object)}
\arguments{\item{\verb{object}}{The \code{\link{GtkCList}} to check.}}
\value{[\code{\link{GtkAdjustment}}] A \code{\link{GtkAdjustment}} object, or NULL if none is currently
being used.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
