\alias{gtkCListNew}
\name{gtkCListNew}
\title{gtkCListNew}
\description{
Creates a new \verb{\link{GtkCList}} widget for use.
\strong{WARNING: \code{gtk_clist_new} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListNew(columns = 1, show = TRUE)}
\arguments{\item{\verb{columns}}{[integer] The number of columns the \verb{\link{GtkCList}} should have.}}
\value{[\verb{\link{GtkWidget}}] A pointer to a new GtkCList object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
