\alias{gtkCListNew}
\name{gtkCListNew}
\title{gtkCListNew}
\description{
Creates a new \code{\link{GtkCList}} widget for use.
\strong{WARNING: \code{gtk_clist_new} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListNew(columns = 1, show = TRUE)}
\arguments{\item{\verb{columns}}{The number of columns the \code{\link{GtkCList}} should have.}}
\value{[\code{\link{GtkWidget}}] A pointer to a new GtkCList object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
