\alias{gtkCListRemove}
\name{gtkCListRemove}
\title{gtkCListRemove}
\description{
Removes the specified row from the CList.
\strong{WARNING: \code{gtk_clist_remove} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListRemove(object, row)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{row}}{The row to remove.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
