\alias{gtkCListRowMove}
\name{gtkCListRowMove}
\title{gtkCListRowMove}
\description{
Allows you to move a row from one position to another in the
list.
\strong{WARNING: \code{gtk_clist_row_move} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListRowMove(object, source.row, dest.row)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{source.row}}{The original position of the row to move.}
\item{\verb{dest.row}}{The position to which the row should be moved.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
