\alias{gtkCListSelectRow}
\name{gtkCListSelectRow}
\title{gtkCListSelectRow}
\description{
Selects the specified row.  Causes the "select-row" signal to be emitted for the specified row and column.
\strong{WARNING: \code{gtk_clist_select_row} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSelectRow(object, row, column)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{row}}{The row to select.}
\item{\verb{column}}{The column to select.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
