\alias{gtkCListSetAutoSort}
\name{gtkCListSetAutoSort}
\title{gtkCListSetAutoSort}
\description{
Turns on or off auto sort of the \code{\link{GtkCList}}. If auto sort is on, then the CList will be resorted when a row is inserted into the CList.
\strong{WARNING: \code{gtk_clist_set_auto_sort} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetAutoSort(object, auto.sort)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{auto.sort}}{whether auto sort should be on or off}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
