\alias{gtkCListSetCellStyle}
\name{gtkCListSetCellStyle}
\title{gtkCListSetCellStyle}
\description{
Sets the style for the specified cell.
\strong{WARNING: \code{gtk_clist_set_cell_style} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetCellStyle(object, row, column, style)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCList}}] The \verb{\link{GtkCList}} to affect.}
\item{\verb{row}}{[integer] The row of the cell.}
\item{\verb{column}}{[integer] The column of the cell.}
\item{\verb{style}}{[\verb{\link{GtkStyle}}] A pointer to a \verb{\link{GtkStyle}} structure.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
