\alias{gtkCListSetCellStyle}
\name{gtkCListSetCellStyle}
\title{gtkCListSetCellStyle}
\description{
Sets the style for the specified cell.
\strong{WARNING: \code{gtk_clist_set_cell_style} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetCellStyle(object, row, column, style)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{row}}{The row of the cell.}
\item{\verb{column}}{The column of the cell.}
\item{\verb{style}}{A pointer to a \code{\link{GtkStyle}} structure.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
