\alias{gtkCListSetColumnJustification}
\name{gtkCListSetColumnJustification}
\title{gtkCListSetColumnJustification}
\description{
Sets the justification to be used for all text in the specified
column.
\strong{WARNING: \code{gtk_clist_set_column_justification} is deprecated and should not be used in newly-written code.  
Sets the justification to be used for all text in the specified
column.}
}
\usage{gtkCListSetColumnJustification(object, column, justification)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCList}}] The \code{\link{GtkCList}} to affect.}
\item{\code{column}}{[integer] The column which should be affected.}
\item{\code{justification}}{[\code{\link{GtkJustification}}] A GtkJustification value for the column.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
