\alias{gtkCListSetColumnJustification}
\name{gtkCListSetColumnJustification}
\title{gtkCListSetColumnJustification}
\description{
Sets the justification to be used for all text in the specified
column.
\strong{WARNING: \code{gtk_clist_set_column_justification} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetColumnJustification(object, column, justification)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{column}}{The column which should be affected.}
\item{\verb{justification}}{A GtkJustification value for the column.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
