\alias{gtkCListSetColumnTitle}
\name{gtkCListSetColumnTitle}
\title{gtkCListSetColumnTitle}
\description{
Sets the title for the specified column.
\strong{WARNING: \code{gtk_clist_set_column_title} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetColumnTitle(object, column, title)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{column}}{The column whose title should be changed.}
\item{\verb{title}}{A string to be the column's title.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
