\alias{gtkCListSetColumnWidth}
\name{gtkCListSetColumnWidth}
\title{gtkCListSetColumnWidth}
\description{
Causes the column specified for the \code{\link{GtkCList}} to be set to
a specified width.
\strong{WARNING: \code{gtk_clist_set_column_width} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetColumnWidth(object, column, width)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{column}}{The column to set the width.}
\item{\verb{width}}{The width, in pixels.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
