\alias{gtkCListSetForeground}
\name{gtkCListSetForeground}
\title{gtkCListSetForeground}
\description{
Sets the foreground color for the specified row.
\strong{WARNING: \code{gtk_clist_set_foreground} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetForeground(object, row, color)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{row}}{The row to affect.}
\item{\verb{color}}{A pointer to a \code{\link{GdkColor}} structure.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
