\alias{gtkCListSetRowHeight}
\name{gtkCListSetRowHeight}
\title{gtkCListSetRowHeight}
\description{
Causes the \code{\link{GtkCList}} to have a specified height for its 
rows.  Setting the row height to 0 allows the \code{\link{GtkCList}} to adjust
automatically to data in the row.
\strong{WARNING: \code{gtk_clist_set_row_height} is deprecated and should not be used in newly-written code.  
Causes the \code{\link{GtkCList}} to have a specified height for its 
rows.  Setting the row height to 0 allows the \code{\link{GtkCList}} to adjust
automatically to data in the row.}
}
\usage{gtkCListSetRowHeight(object, height)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCList}}] The \code{\link{GtkCList}} to affect.}
\item{\code{height}}{[numeric] The height, in pixels.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
