\alias{gtkCListSetRowStyle}
\name{gtkCListSetRowStyle}
\title{gtkCListSetRowStyle}
\description{
Sets the style for all cells in the specified row.
\strong{WARNING: \code{gtk_clist_set_row_style} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetRowStyle(object, row, style)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCList}}] The \code{\link{GtkCList}} to affect.}
\item{\code{row}}{[integer] The row to affect.}
\item{\code{style}}{[\code{\link{GtkStyle}}] A pointer to a \code{\link{GtkStyle}} to set.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
