\alias{gtkCListSetSelectable}
\name{gtkCListSetSelectable}
\title{gtkCListSetSelectable}
\description{
Sets whether the specified row is selectable or not.
\strong{WARNING: \code{gtk_clist_set_selectable} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetSelectable(object, row, selectable)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{row}}{The row to affect.}
\item{\verb{selectable}}{\code{TRUE} or \code{FALSE}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
