\alias{gtkCListSetShift}
\name{gtkCListSetShift}
\title{gtkCListSetShift}
\description{
Sets the vertical and horizontal shift of the specified cell.
\strong{WARNING: \code{gtk_clist_set_shift} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetShift(object, row, column, vertical, horizontal)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{row}}{The row of the cell.}
\item{\verb{column}}{The column of the cell.}
\item{\verb{vertical}}{The value to set for the vertical shift.}
\item{\verb{horizontal}}{The value to set for the vertical shift.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
