\alias{gtkCListSetSortType}
\name{gtkCListSetSortType}
\title{gtkCListSetSortType}
\description{
Sets the sort type of the \verb{GtkClist}. This is either GTK_SORT_ASCENDING for
ascening sort or GTK_SORT_DESCENDING for descending sort.
\strong{WARNING: \code{gtk_clist_set_sort_type} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetSortType(object, sort.type)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{sort.type}}{the \code{\link{GtkSortType}} to use}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
