\alias{gtkCListSetText}
\name{gtkCListSetText}
\title{gtkCListSetText}
\description{
Sets the displayed text in the specified cell.
\strong{WARNING: \code{gtk_clist_set_text} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetText(w, row, cols, values, zeroBased = TRUE)}
\arguments{\item{\verb{row}}{The row of the cell.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
