\alias{gtkCListSetUseDragIcons}
\name{gtkCListSetUseDragIcons}
\title{gtkCListSetUseDragIcons}
\description{
Determines whether the \verb{GtkClist} should use icons when
doing drag-and-drop operations.
\strong{WARNING: \code{gtk_clist_set_use_drag_icons} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetUseDragIcons(object, use.icons)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{use.icons}}{\code{TRUE} or \code{FALSE}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
