\alias{gtkCListSetVadjustment}
\name{gtkCListSetVadjustment}
\title{gtkCListSetVadjustment}
\description{
Allows you to set the \code{\link{GtkAdjustment}} to be used for the vertical
aspect of the \code{\link{GtkCList}} widget.
\strong{WARNING: \code{gtk_clist_set_vadjustment} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSetVadjustment(object, adjustment)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{adjustment}}{A pointer to a \code{\link{GtkAdjustment}} widget, or NULL.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
