\alias{gtkCListSort}
\name{gtkCListSort}
\title{gtkCListSort}
\description{
Sorts the \verb{GtkClist} according to the current compare function, which
can be set with the \code{\link{gtkCListSetCompareFunc}} function.
\strong{WARNING: \code{gtk_clist_sort} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSort(object)}
\arguments{\item{\verb{object}}{The \code{\link{GtkCList}} to sort.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
