\alias{gtkCListSwapRows}
\name{gtkCListSwapRows}
\title{gtkCListSwapRows}
\description{
Swaps the two specified rows with each other.
\strong{WARNING: \code{gtk_clist_swap_rows} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCListSwapRows(object, row1, row2)}
\arguments{
\item{\verb{object}}{The \code{\link{GtkCList}} to affect.}
\item{\verb{row1}}{Number of the first row.}
\item{\verb{row2}}{Number of the second row.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
