\alias{gtkColorButtonGetColor}
\name{gtkColorButtonGetColor}
\title{gtkColorButtonGetColor}
\description{Sets \code{color} to be the current color in the \code{\link{GtkColorButton}} widget.}
\usage{gtkColorButtonGetColor(object, color)}
\arguments{
\item{\code{object}}{[\code{\link{GtkColorButton}}]  a \code{\link{GtkColorButton}}.}
\item{\code{color}}{[\code{\link{GdkColor}}]  a \code{\link{GdkColor}} to fill in with the current color.}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
