\alias{gtkColorButtonGetColor}
\name{gtkColorButtonGetColor}
\title{gtkColorButtonGetColor}
\description{Sets \code{color} to be the current color in the \code{\link{GtkColorButton}} widget.}
\usage{gtkColorButtonGetColor(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkColorButton}}.}}
\details{Since 2.4}
\value{
A list containing the following elements:
\item{\verb{color}}{a \code{\link{GdkColor}} to fill in with the current color.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
