\alias{gtkColorButtonSetAlpha}
\name{gtkColorButtonSetAlpha}
\title{gtkColorButtonSetAlpha}
\description{Sets the current opacity to be \code{alpha}.}
\usage{gtkColorButtonSetAlpha(object, alpha)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkColorButton}}]  a \verb{\link{GtkColorButton}}.}
\item{\verb{alpha}}{[integer]  an integer between 0 and 65535.}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
