\alias{gtkColorButtonSetColor}
\name{gtkColorButtonSetColor}
\title{gtkColorButtonSetColor}
\description{Sets the current color to be \code{color}.}
\usage{gtkColorButtonSetColor(object, color)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkColorButton}}]  a \verb{\link{GtkColorButton}}.}
\item{\verb{color}}{[\verb{\link{GdkColor}}]  A \verb{\link{GdkColor}} to set the current color with.}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
