\alias{gtkColorButtonSetTitle}
\name{gtkColorButtonSetTitle}
\title{gtkColorButtonSetTitle}
\description{Sets the title for the color selection dialog.}
\usage{gtkColorButtonSetTitle(object, title)}
\arguments{
\item{\code{object}}{[\code{\link{GtkColorButton}}]  a \code{\link{GtkColorButton}}}
\item{\code{title}}{[character]  String containing new window title.}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
