\alias{gtkColorButtonSetUseAlpha}
\name{gtkColorButtonSetUseAlpha}
\title{gtkColorButtonSetUseAlpha}
\description{Sets whether or not the color button should use the alpha channel.}
\usage{gtkColorButtonSetUseAlpha(object, use.alpha)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkColorButton}}.}
\item{\verb{use.alpha}}{\code{TRUE} if color button should use alpha channel, \code{FALSE} if not.}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
