\alias{gtkColorSelectionDialogNew}
\name{gtkColorSelectionDialogNew}
\title{gtkColorSelectionDialogNew}
\description{Creates a new \code{\link{GtkColorSelectionDialog}}.}
\usage{gtkColorSelectionDialogNew(title = NULL, show = TRUE)}
\arguments{\item{\verb{title}}{a string containing the title text for the dialog.}}
\value{[\code{\link{GtkWidget}}] a \code{\link{GtkColorSelectionDialog}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
