\alias{gtkColorSelectionGetCurrentColor}
\name{gtkColorSelectionGetCurrentColor}
\title{gtkColorSelectionGetCurrentColor}
\description{Sets \code{color} to be the current color in the GtkColorSelection widget.}
\usage{gtkColorSelectionGetCurrentColor(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkColorSelection}}.}}
\value{
A list containing the following elements:
\item{\verb{color}}{a \code{\link{GdkColor}} to fill in with the current color. \emph{[  \acronym{out}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
