\alias{gtkColorSelectionGetHasOpacityControl}
\name{gtkColorSelectionGetHasOpacityControl}
\title{gtkColorSelectionGetHasOpacityControl}
\description{Determines whether the colorsel has an opacity control.}
\usage{gtkColorSelectionGetHasOpacityControl(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkColorSelection}}.}}
\value{[logical] \code{TRUE} if the \code{colorsel} has an opacity control.  \code{FALSE} if it does't.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
