\alias{gtkColorSelectionGetHasPalette}
\name{gtkColorSelectionGetHasPalette}
\title{gtkColorSelectionGetHasPalette}
\description{Determines whether the color selector has a color palette.}
\usage{gtkColorSelectionGetHasPalette(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkColorSelection}}.}}
\value{[logical] \code{TRUE} if the selector has a palette.  \code{FALSE} if it hasn't.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
