\alias{gtkColorSelectionGetPreviousColor}
\name{gtkColorSelectionGetPreviousColor}
\title{gtkColorSelectionGetPreviousColor}
\description{Fills \code{color} in with the original color value.}
\usage{gtkColorSelectionGetPreviousColor(object, color)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkColorSelection}}.}
\item{\verb{color}}{a \code{\link{GdkColor}} to fill in with the original color value.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
