\alias{gtkColorSelectionIsAdjusting}
\name{gtkColorSelectionIsAdjusting}
\title{gtkColorSelectionIsAdjusting}
\description{Gets the current state of the \code{colorsel}.}
\usage{gtkColorSelectionIsAdjusting(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkColorSelection}}.}}
\value{[logical] \code{TRUE} if the user is currently dragging a color around, and \code{FALSE}
if the selection has stopped.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
