\alias{gtkColorSelectionNew}
\name{gtkColorSelectionNew}
\title{gtkColorSelectionNew}
\description{Creates a new GtkColorSelection.}
\usage{gtkColorSelectionNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}]  a new \verb{\link{GtkColorSelection}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
