\alias{gtkColorSelectionSetCurrentColor}
\name{gtkColorSelectionSetCurrentColor}
\title{gtkColorSelectionSetCurrentColor}
\description{Sets the current color to be \code{color}.  The first time this is called, it will
also set the original color to be \code{color} too.}
\usage{gtkColorSelectionSetCurrentColor(object, color)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkColorSelection}}.}
\item{\verb{color}}{A \code{\link{GdkColor}} to set the current color with.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
