\alias{gtkColorSelectionSetHasOpacityControl}
\name{gtkColorSelectionSetHasOpacityControl}
\title{gtkColorSelectionSetHasOpacityControl}
\description{Sets the \code{colorsel} to use or not use opacity.}
\usage{gtkColorSelectionSetHasOpacityControl(object, has.opacity)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkColorSelection}}.}
\item{\verb{has.opacity}}{\code{TRUE} if \code{colorsel} can set the opacity, \code{FALSE} otherwise.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
