\alias{gtkColorSelectionSetHasPalette}
\name{gtkColorSelectionSetHasPalette}
\title{gtkColorSelectionSetHasPalette}
\description{Shows and hides the palette based upon the value of \code{has.palette}.}
\usage{gtkColorSelectionSetHasPalette(object, has.palette)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkColorSelection}}.}
\item{\verb{has.palette}}{\code{TRUE} if palette is to be visible, \code{FALSE} otherwise.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
