\alias{gtkComboBoxEntryNew}
\name{gtkComboBoxEntryNew}
\title{gtkComboBoxEntryNew}
\description{Creates a new \code{\link{GtkComboBoxEntry}} which has a \code{\link{GtkEntry}} as child. After
construction, you should set a model using \code{\link{gtkComboBoxSetModel}} and a
text column using \code{\link{gtkComboBoxEntrySetTextColumn}}.}
\usage{gtkComboBoxEntryNew(show = TRUE)}
\details{Since 2.4}
\value{[\code{\link{GtkWidget}}] A new \code{\link{GtkComboBoxEntry}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
