\alias{gtkComboBoxEntrySetTextColumn}
\name{gtkComboBoxEntrySetTextColumn}
\title{gtkComboBoxEntrySetTextColumn}
\description{Sets the model column which \code{entry.box} should use to get strings from
to be \code{text.column}.}
\usage{gtkComboBoxEntrySetTextColumn(object, text.column)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkComboBoxEntry}}.}
\item{\verb{text.column}}{A column in \code{model} to get the strings from.}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
