\alias{gtkComboBoxGetActiveIter}
\name{gtkComboBoxGetActiveIter}
\title{gtkComboBoxGetActiveIter}
\description{Sets \code{iter} to point to the current active item, if it exists.}
\usage{gtkComboBoxGetActiveIter(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkComboBox}}}}
\details{Since 2.4}
\value{
A list containing the following elements:
\item{retval}{[logical] \code{TRUE}, if \code{iter} was set}
\item{\verb{iter}}{The uninitialized \code{\link{GtkTreeIter}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
