\alias{gtkComboBoxGetModel}
\name{gtkComboBoxGetModel}
\title{gtkComboBoxGetModel}
\description{Returns the \code{\link{GtkTreeModel}} which is acting as data source for \code{combo.box}.}
\usage{gtkComboBoxGetModel(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkComboBox}}}}
\details{Since 2.4}
\value{[\code{\link{GtkTreeModel}}] A \code{\link{GtkTreeModel}} which was passed during construction. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
