\alias{gtkComboBoxGetRowSeparatorFunc}
\name{gtkComboBoxGetRowSeparatorFunc}
\title{gtkComboBoxGetRowSeparatorFunc}
\description{Returns the current row separator function.}
\usage{gtkComboBoxGetRowSeparatorFunc(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkComboBox}}}}
\details{Since 2.6}
\value{[\code{\link{GtkTreeViewRowSeparatorFunc}}] the current row separator function.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
