\alias{gtkComboBoxGetRowSpanColumn}
\name{gtkComboBoxGetRowSpanColumn}
\title{gtkComboBoxGetRowSpanColumn}
\description{Returns the column with row span information for \code{combo.box}.}
\usage{gtkComboBoxGetRowSpanColumn(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkComboBox}}}}
\details{Since 2.6}
\value{[integer] the row span column.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
