\alias{gtkComboBoxNew}
\name{gtkComboBoxNew}
\title{gtkComboBoxNew}
\description{Creates a new empty \code{\link{GtkComboBox}}.}
\usage{gtkComboBoxNew(show = TRUE)}
\details{Since 2.4}
\value{[\code{\link{GtkWidget}}] A new \code{\link{GtkComboBox}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
