\alias{gtkComboBoxNewWithModel}
\name{gtkComboBoxNewWithModel}
\title{gtkComboBoxNewWithModel}
\description{Creates a new \code{\link{GtkComboBox}} with the model initialized to \code{model}.}
\usage{gtkComboBoxNewWithModel(model, show = TRUE)}
\arguments{\item{\verb{model}}{A \code{\link{GtkTreeModel}}.}}
\details{Since 2.4}
\value{[\code{\link{GtkWidget}}] A new \code{\link{GtkComboBox}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
