\alias{gtkComboBoxSetActive}
\name{gtkComboBoxSetActive}
\title{gtkComboBoxSetActive}
\description{Sets the active item of \code{combo.box} to be the item at \code{index}.}
\usage{gtkComboBoxSetActive(object, index)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkComboBox}}}
\item{\verb{index}}{An index in the model passed during construction, or -1 to have
no active item}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
