\alias{gtkComboBoxSetActiveIter}
\name{gtkComboBoxSetActiveIter}
\title{gtkComboBoxSetActiveIter}
\description{Sets the current active item to be the one referenced by \code{iter}. 
\code{iter} must correspond to a path of depth one.}
\usage{gtkComboBoxSetActiveIter(object, iter)}
\arguments{
\item{\code{object}}{[\code{\link{GtkComboBox}}]  A \code{\link{GtkComboBox}}}
\item{\code{iter}}{[\code{\link{GtkTreeIter}}]  The \code{\link{GtkTreeIter}}}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
