\alias{gtkComboBoxSetFocusOnClick}
\name{gtkComboBoxSetFocusOnClick}
\title{gtkComboBoxSetFocusOnClick}
\description{Sets whether the combo box will grab focus when it is clicked with 
the mouse. Making mouse clicks not grab focus is useful in places 
like toolbars where you don't want the keyboard focus removed from 
the main area of the application.}
\usage{gtkComboBoxSetFocusOnClick(object, focus.on.click)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkComboBox}}}
\item{\verb{focus.on.click}}{whether the combo box grabs focus when clicked
with the mouse}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
