\alias{gtkComboBoxSetRowSpanColumn}
\name{gtkComboBoxSetRowSpanColumn}
\title{gtkComboBoxSetRowSpanColumn}
\description{Sets the column with row span information for \code{combo.box} to be \code{row.span}.
The row span column contains integers which indicate how many rows
an item should span.}
\usage{gtkComboBoxSetRowSpanColumn(object, row.span)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkComboBox}}.}
\item{\verb{row.span}}{A column in the model passed during construction.}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
