\alias{gtkComboNew}
\name{gtkComboNew}
\title{gtkComboNew}
\description{
Creates a new \code{\link{GtkCombo}}.
\strong{WARNING: \code{gtk_combo_new} has been deprecated since version 2.4 and should not be used in newly-written code. Use \code{\link{GtkComboBox}} instead.}
}
\usage{gtkComboNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkCombo}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
