\alias{gtkComboSetUseArrows}
\name{gtkComboSetUseArrows}
\title{gtkComboSetUseArrows}
\description{
Specifies if the arrow (cursor) keys can be used to step through the items in
the list. This is on by default.
\strong{WARNING: \code{gtk_combo_set_use_arrows} has been deprecated since version 2.4 and should not be used in newly-written code. Use \code{\link{GtkComboBox}} instead.}
}
\usage{gtkComboSetUseArrows(object, val)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkCombo}}.}
\item{\verb{val}}{\code{TRUE} if the arrow keys can be used to step through the items in
the list.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
