\alias{gtkComboSetUseArrowsAlways}
\name{gtkComboSetUseArrowsAlways}
\title{gtkComboSetUseArrowsAlways}
\description{
Obsolete function, does nothing.
\strong{WARNING: \code{gtk_combo_set_use_arrows_always} has been deprecated since version 2.4 and should not be used in newly-written code. Use \code{\link{GtkComboBox}} instead.}
}
\usage{gtkComboSetUseArrowsAlways(object, val)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkCombo}}.}
\item{\verb{val}}{unused}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
