\alias{gtkContainerAddWithProperties}
\name{gtkContainerAddWithProperties}
\title{gtkContainerAddWithProperties}
\description{Adds \code{widget} to \code{container}, setting child properties at the same time.
See \code{\link{gtkContainerAdd}} and \code{\link{gtkContainerChildSet}} for more details.}
\usage{gtkContainerAddWithProperties(object, widget, ...)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkContainer}}}
\item{\verb{widget}}{a widget to be placed inside \code{container}}
\item{\verb{...}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
