\alias{gtkContainerCheckResize}
\name{gtkContainerCheckResize}
\title{gtkContainerCheckResize}
\description{\emph{undocumented
}}
\usage{gtkContainerCheckResize(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkContainer}}] }}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
