\alias{gtkContainerChildGet}
\name{gtkContainerChildGet}
\title{gtkContainerChildGet}
\description{Gets the values of one or more child properties for \code{child} and \code{container}.}
\usage{gtkContainerChildGet(object, child, ...)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkContainer}}]  a \verb{\link{GtkContainer}}}
\item{\verb{child}}{[\verb{\link{GtkWidget}}]  a widget which is a child of \code{container}}
\item{\verb{...}}{ a list of property names and \verb{R object}*, 
          starting with \code{first.prop.name}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
