\alias{gtkContainerChildGet}
\name{gtkContainerChildGet}
\title{gtkContainerChildGet}
\description{Gets the values of one or more child properties for \code{child} and \code{container}.}
\usage{gtkContainerChildGet(object, child, ...)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkContainer}}}
\item{\verb{child}}{a widget which is a child of \code{container}}
\item{\verb{...}}{return location for the first property, followed
optionally by more name/return location pairs, followed by \code{NULL}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
