\alias{gtkContainerChildGet}
\name{gtkContainerChildGet}
\title{gtkContainerChildGet}
\description{Gets the values of one or more child properties for \code{child} and \code{container}.}
\usage{gtkContainerChildGet(object, child, ...)}
\arguments{
\item{\code{object}}{[\code{\link{GtkContainer}}]  a \code{\link{GtkContainer}}}
\item{\code{child}}{[\code{\link{GtkWidget}}]  a widget which is a child of \code{container}}
\item{\code{...}}{ a list of property names and \code{R object}*, 
          starting with \code{first.prop.name}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
