\alias{gtkContainerChildGetProperty}
\name{gtkContainerChildGetProperty}
\title{gtkContainerChildGetProperty}
\description{Gets the value of a child property for \code{child} and \code{container}.}
\usage{gtkContainerChildGetProperty(object, child, property.name)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkContainer}}}
\item{\verb{child}}{a widget which is a child of \code{container}}
\item{\verb{property.name}}{the name of the property to get}
}
\value{
A list containing the following elements:
\item{\verb{value}}{a location to return the value}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
