\alias{gtkContainerChildSet}
\name{gtkContainerChildSet}
\title{gtkContainerChildSet}
\description{Sets one or more child properties for \code{child} and \code{container}.}
\usage{gtkContainerChildSet(object, child, ...)}
\arguments{\item{\code{...}}{[\code{\link{GtkWidget}}]  a widget which is a child of \code{container}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
