\alias{gtkContainerChildSet}
\name{gtkContainerChildSet}
\title{gtkContainerChildSet}
\description{Sets one or more child properties for \code{child} and \code{container}.}
\usage{gtkContainerChildSet(object, child, ...)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkContainer}}}
\item{\verb{child}}{a widget which is a child of \code{container}}
\item{\verb{...}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
