\alias{gtkContainerChildSetProperty}
\name{gtkContainerChildSetProperty}
\title{gtkContainerChildSetProperty}
\description{Sets a child property for \code{child} and \code{container}.}
\usage{gtkContainerChildSetProperty(object, child, property.name, value)}
\arguments{
\item{\code{object}}{[\code{\link{GtkContainer}}]  a \code{\link{GtkContainer}}}
\item{\code{child}}{[\code{\link{GtkWidget}}]  a widget which is a child of \code{container}}
\item{\code{property.name}}{[character]  the name of the property to set}
\item{\code{value}}{[R object]  the value to set the property to}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
