\alias{gtkContainerChildSetProperty}
\name{gtkContainerChildSetProperty}
\title{gtkContainerChildSetProperty}
\description{Sets a child property for \code{child} and \code{container}.}
\usage{gtkContainerChildSetProperty(object, child, property.name, value)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkContainer}}}
\item{\verb{child}}{a widget which is a child of \code{container}}
\item{\verb{property.name}}{the name of the property to set}
\item{\verb{value}}{the value to set the property to}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
