\alias{gtkContainerClassInstallChildProperty}
\name{gtkContainerClassInstallChildProperty}
\title{gtkContainerClassInstallChildProperty}
\description{Installs a child property on a container class.}
\usage{gtkContainerClassInstallChildProperty(cclass, property.id, pspec)}
\arguments{
\item{\verb{cclass}}{[\verb{\link{GtkContainerClass}}]  a \verb{\link{GtkContainerClass}}}
\item{\verb{property.id}}{[numeric]  the id for the property}
\item{\verb{pspec}}{[\verb{\link{GParamSpec}}]  the \verb{\link{GParamSpec}} for the property}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
