\alias{gtkContainerClassInstallChildProperty}
\name{gtkContainerClassInstallChildProperty}
\title{gtkContainerClassInstallChildProperty}
\description{Installs a child property on a container class.}
\usage{gtkContainerClassInstallChildProperty(cclass, property.id, pspec)}
\arguments{
\item{\verb{cclass}}{a \code{\link{GtkContainerClass}}}
\item{\verb{property.id}}{the id for the property}
\item{\verb{pspec}}{the \code{\link{GParamSpec}} for the property}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
