\alias{gtkContainerGetChildren}
\name{gtkContainerGetChildren}
\title{gtkContainerGetChildren}
\description{Returns the container's non-internal children. See
\code{\link{gtkContainerForall}} for details on what constitutes an "internal" child.}
\usage{gtkContainerGetChildren(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkContainer}}]  a \verb{\link{GtkContainer}}}}
\value{[list]  a newly-allocated list of the container's non-internal children.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
