\alias{gtkContainerGetFocusHadjustment}
\name{gtkContainerGetFocusHadjustment}
\title{gtkContainerGetFocusHadjustment}
\description{Retrieves the horizontal focus adjustment for the container. See
\code{\link{gtkContainerSetFocusHadjustment}}.}
\usage{gtkContainerGetFocusHadjustment(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkContainer}}]  a \code{\link{GtkContainer}}}}
\value{[\code{\link{GtkAdjustment}}]  the horizontal focus adjustment, or \code{NULL} if
  none has been set.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
