\alias{gtkContainerGetFocusHadjustment}
\name{gtkContainerGetFocusHadjustment}
\title{gtkContainerGetFocusHadjustment}
\description{Retrieves the horizontal focus adjustment for the container. See
\code{\link{gtkContainerSetFocusHadjustment}}.}
\usage{gtkContainerGetFocusHadjustment(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkContainer}}}}
\value{[\code{\link{GtkAdjustment}}] the horizontal focus adjustment, or \code{NULL} if
none has been set. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
