\alias{gtkContainerGetFocusVadjustment}
\name{gtkContainerGetFocusVadjustment}
\title{gtkContainerGetFocusVadjustment}
\description{Retrieves the vertical focus adjustment for the container. See
\code{\link{gtkContainerSetFocusVadjustment}}.}
\usage{gtkContainerGetFocusVadjustment(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkContainer}}]  a \verb{\link{GtkContainer}}}}
\value{[\verb{\link{GtkAdjustment}}]  the vertical focus adjustment, or \code{NULL} if
  none has been set.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
