\alias{gtkContainerResizeChildren}
\name{gtkContainerResizeChildren}
\title{gtkContainerResizeChildren}
\description{\emph{undocumented
}}
\usage{gtkContainerResizeChildren(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkContainer}}] }}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
