\alias{gtkContainerSetFocusChild}
\name{gtkContainerSetFocusChild}
\title{gtkContainerSetFocusChild}
\description{\emph{undocumented
}}
\usage{gtkContainerSetFocusChild(object, child)}
\arguments{
\item{\code{object}}{[\code{\link{GtkContainer}}] }
\item{\code{child}}{[\code{\link{GtkWidget}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
