\alias{gtkContainerSetFocusChild}
\name{gtkContainerSetFocusChild}
\title{gtkContainerSetFocusChild}
\description{\emph{undocumented
}}
\usage{gtkContainerSetFocusChild(object, child)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkContainer}}] }
\item{\verb{child}}{[\verb{\link{GtkWidget}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
