\alias{gtkCTreeCollapse}
\name{gtkCTreeCollapse}
\title{gtkCTreeCollapse}
\description{
Collapse one node.
\strong{WARNING: \code{gtk_ctree_collapse} is deprecated and should not be used in newly-written code.  
Collapse one node.}
}
\usage{gtkCTreeCollapse(object, node)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
