\alias{gtkCTreeCollapseRecursive}
\name{gtkCTreeCollapseRecursive}
\title{gtkCTreeCollapseRecursive}
\description{
Collapse one node and all its subnodes.
\strong{WARNING: \code{gtk_ctree_collapse_recursive} is deprecated and should not be used in newly-written code.  
Collapse one node and all its subnodes.}
}
\usage{gtkCTreeCollapseRecursive(object, node)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
