\alias{gtkCTreeCollapseToDepth}
\name{gtkCTreeCollapseToDepth}
\title{gtkCTreeCollapseToDepth}
\description{
Collapse a node and its children up to the depth given.
\strong{WARNING: \code{gtk_ctree_collapse_to_depth} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeCollapseToDepth(object, node, depth)}
\arguments{
\item{\verb{object}}{The (absolute) depth up to which to collapse nodes.}
\item{\verb{node}}{\emph{undocumented
}}
\item{\verb{depth}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
