\alias{gtkCTreeExpand}
\name{gtkCTreeExpand}
\title{gtkCTreeExpand}
\description{
Expand one node.
\strong{WARNING: \code{gtk_ctree_expand} is deprecated and should not be used in newly-written code.  
Expand one node.}
}
\usage{gtkCTreeExpand(object, node)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
