\alias{gtkCTreeExpandRecursive}
\name{gtkCTreeExpandRecursive}
\title{gtkCTreeExpandRecursive}
\description{
Expand one node and all nodes underneath.
\strong{WARNING: \code{gtk_ctree_expand_recursive} is deprecated and should not be used in newly-written code.  
Expand one node and all nodes underneath.}
}
\usage{gtkCTreeExpandRecursive(object, node)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
