\alias{gtkCTreeExpandRecursive}
\name{gtkCTreeExpandRecursive}
\title{gtkCTreeExpandRecursive}
\description{
Expand one node and all nodes underneath.
\strong{WARNING: \code{gtk_ctree_expand_recursive} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeExpandRecursive(object, node)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{node}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
