\alias{gtkCTreeExpandToDepth}
\name{gtkCTreeExpandToDepth}
\title{gtkCTreeExpandToDepth}
\description{
Expand a node and its children up to the depth given.
\strong{WARNING: \code{gtk_ctree_expand_to_depth} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeExpandToDepth(object, node, depth)}
\arguments{
\item{\verb{object}}{The (absolute) depth up to which to expand nodes.}
\item{\verb{node}}{\emph{undocumented
}}
\item{\verb{depth}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
