\alias{gtkCTreeExportToGnode}
\name{gtkCTreeExportToGnode}
\title{gtkCTreeExportToGnode}
\description{
FIXME
\strong{WARNING: \code{gtk_ctree_export_to_gnode} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeExportToGnode(object, parent, sibling, node, func, data = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCTree}}] }
\item{\verb{parent}}{[GNode] }
\item{\verb{sibling}}{[GNode] }
\item{\verb{node}}{[\verb{\link{GtkCTreeNode}}] }
\item{\verb{func}}{[\verb{\link{GtkCTreeGNodeFunc}}] }
\item{\verb{data}}{[R object] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
