\alias{gtkCTreeExportToGnode}
\name{gtkCTreeExportToGnode}
\title{gtkCTreeExportToGnode}
\description{
FIXME
\strong{WARNING: \code{gtk_ctree_export_to_gnode} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeExportToGnode(object, parent, sibling, node, func, data = NULL)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{parent}}{\emph{undocumented
}}
\item{\verb{sibling}}{\emph{undocumented
}}
\item{\verb{node}}{\emph{undocumented
}}
\item{\verb{func}}{\emph{undocumented
}}
\item{\verb{data}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
