\alias{gtkCTreeFindAllByRowData}
\name{gtkCTreeFindAllByRowData}
\title{gtkCTreeFindAllByRowData}
\description{
Finds all nodes in the tree under \code{node} that have the given user data pointer.
\strong{WARNING: \code{gtk_ctree_find_all_by_row_data} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeFindAllByRowData(object, node, data = NULL)}
\arguments{
\item{\verb{object}}{A list of nodes that have the given data pointer.}
\item{\verb{node}}{\emph{undocumented
}}
\item{\verb{data}}{\emph{undocumented
}}
}
\value{[list] A list of nodes that have the given data pointer.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
