\alias{gtkCTreeGetNodeInfo}
\name{gtkCTreeGetNodeInfo}
\title{gtkCTreeGetNodeInfo}
\description{
Get information corresponding to a node.
Any of the return parameters can be null.
\strong{WARNING: \code{gtk_ctree_get_node_info} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeGetNodeInfo(object, node)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{node}}{\emph{undocumented
}}
}
\value{
A list containing the following elements:
\item{\verb{text}}{\emph{undocumented
}}
\item{\verb{spacing}}{\emph{undocumented
}}
\item{\verb{pixmap.closed}}{\emph{undocumented
}}
\item{\verb{mask.closed}}{\emph{undocumented
}}
\item{\verb{pixmap.opened}}{\emph{undocumented
}}
\item{\verb{mask.opened}}{\emph{undocumented
}}
\item{\verb{is.leaf}}{\emph{undocumented
}}
\item{\verb{expanded}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
