\alias{gtkCTreeInsertGnode}
\name{gtkCTreeInsertGnode}
\title{gtkCTreeInsertGnode}
\description{
FIXME
\strong{WARNING: \code{gtk_ctree_insert_gnode} is deprecated and should not be used in newly-written code.  
FIXME}
}
\usage{gtkCTreeInsertGnode(object, parent, sibling, gnode, func, data = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{parent}}{[\code{\link{GtkCTreeNode}}] }
\item{\code{sibling}}{[\code{\link{GtkCTreeNode}}] }
\item{\code{gnode}}{[GNode] }
\item{\code{func}}{[\code{\link{GtkCTreeGNodeFunc}}] }
\item{\code{data}}{[R object] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
