\alias{gtkCTreeInsertGnode}
\name{gtkCTreeInsertGnode}
\title{gtkCTreeInsertGnode}
\description{
FIXME
\strong{WARNING: \code{gtk_ctree_insert_gnode} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeInsertGnode(object, parent, sibling, gnode, func, data = NULL)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{parent}}{\emph{undocumented
}}
\item{\verb{sibling}}{\emph{undocumented
}}
\item{\verb{gnode}}{\emph{undocumented
}}
\item{\verb{func}}{\emph{undocumented
}}
\item{\verb{data}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
