\alias{gtkCTreeIsAncestor}
\name{gtkCTreeIsAncestor}
\title{gtkCTreeIsAncestor}
\description{

\strong{WARNING: \code{gtk_ctree_is_ancestor} is deprecated and should not be used in newly-written code.  }
}
\usage{gtkCTreeIsAncestor(object, node, child)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
\item{\code{child}}{[\code{\link{GtkCTreeNode}}] }
}
\value{[logical] True is \code{node} is an ancestor of \code{child}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
