\alias{gtkCTreeIsAncestor}
\name{gtkCTreeIsAncestor}
\title{gtkCTreeIsAncestor}
\description{

\strong{WARNING: \code{gtk_ctree_is_ancestor} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeIsAncestor(object, node, child)}
\arguments{
\item{\verb{object}}{True is \code{node} is an ancestor of \code{child}.}
\item{\verb{node}}{\emph{undocumented
}}
\item{\verb{child}}{\emph{undocumented
}}
}
\value{[logical] True is \code{node} is an ancestor of \code{child}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
