\alias{gtkCTreeIsHotSpot}
\name{gtkCTreeIsHotSpot}
\title{gtkCTreeIsHotSpot}
\description{

\strong{WARNING: \code{gtk_ctree_is_hot_spot} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeIsHotSpot(object, x, y)}
\arguments{
\item{\verb{object}}{True if the given coordinates lie on an expander button.}
\item{\verb{x}}{\emph{undocumented
}}
\item{\verb{y}}{\emph{undocumented
}}
}
\value{[logical] True if the given coordinates lie on an expander button.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
