\alias{gtkCTreeLast}
\name{gtkCTreeLast}
\title{gtkCTreeLast}
\description{
Returns the last child of the last child of the last child...
of the given node.
\strong{WARNING: \code{gtk_ctree_last} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeLast(object, node)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{node}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
