\alias{gtkCTreeMove}
\name{gtkCTreeMove}
\title{gtkCTreeMove}
\description{
Move a node in the tree to another location.
\strong{WARNING: \code{gtk_ctree_move} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeMove(object, node, new.parent = NULL, new.sibling = NULL)}
\arguments{
\item{\verb{object}}{The node to be moved.}
\item{\verb{node}}{. \emph{[  \acronym{allow-none}  ]}}
\item{\verb{new.parent}}{. \emph{[  \acronym{allow-none}  ]}}
\item{\verb{new.sibling}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
