\alias{gtkCTreeNew}
\name{gtkCTreeNew}
\title{gtkCTreeNew}
\description{
Create a new \code{\link{GtkCTree}} widget.
\strong{WARNING: \code{gtk_ctree_new} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNew(columns = 1, tree.column = 0, show = TRUE)}
\arguments{
\item{\verb{columns}}{Number of columns.}
\item{\verb{tree.column}}{Which columns has the tree graphic.}
}
\value{[\code{\link{GtkWidget}}] The new \code{\link{GtkCTree}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
