\alias{gtkCTreeNodeGetCellStyle}
\name{gtkCTreeNodeGetCellStyle}
\title{gtkCTreeNodeGetCellStyle}
\description{
Get the style of an individual cell.
\strong{WARNING: \code{gtk_ctree_node_get_cell_style} is deprecated and should not be used in newly-written code.  
Get the style of an individual cell.}
}
\usage{gtkCTreeNodeGetCellStyle(object, node, column)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
\item{\code{column}}{[integer] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
