\alias{gtkCTreeNodeGetCellStyle}
\name{gtkCTreeNodeGetCellStyle}
\title{gtkCTreeNodeGetCellStyle}
\description{
Get the style of an individual cell.
\strong{WARNING: \code{gtk_ctree_node_get_cell_style} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeGetCellStyle(object, node, column)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{node}}{\emph{undocumented
}}
\item{\verb{column}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
