\alias{gtkCTreeNodeGetCellType}
\name{gtkCTreeNodeGetCellType}
\title{gtkCTreeNodeGetCellType}
\description{

\strong{WARNING: \code{gtk_ctree_node_get_cell_type} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeGetCellType(object, node, column)}
\arguments{
\item{\verb{object}}{The type of the given cell.}
\item{\verb{node}}{\emph{undocumented
}}
\item{\verb{column}}{\emph{undocumented
}}
}
\value{[\code{\link{GtkCellType}}] The type of the given cell.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
