\alias{gtkCTreeNodeGetPixtext}
\name{gtkCTreeNodeGetPixtext}
\title{gtkCTreeNodeGetPixtext}
\description{
Get the parameters of a cell containing both a pixmap and text.
\strong{WARNING: \code{gtk_ctree_node_get_pixtext} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeGetPixtext(object, node, column)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{node}}{\emph{undocumented
}}
\item{\verb{column}}{\emph{undocumented
}}
}
\value{
A list containing the following elements:
\item{\verb{text}}{\emph{undocumented
}}
\item{\verb{spacing}}{\emph{undocumented
}}
\item{\verb{pixmap}}{\emph{undocumented
}}
\item{\verb{mask}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
