\alias{gtkCTreeNodeGetRowData}
\name{gtkCTreeNodeGetRowData}
\title{gtkCTreeNodeGetRowData}
\description{

\strong{WARNING: \code{gtk_ctree_node_get_row_data} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeGetRowData(object, node)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{node}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
