\alias{gtkCTreeNodeGetRowStyle}
\name{gtkCTreeNodeGetRowStyle}
\title{gtkCTreeNodeGetRowStyle}
\description{
Get the style of a row.
\strong{WARNING: \code{gtk_ctree_node_get_row_style} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeGetRowStyle(object, node)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCTree}}] }
\item{\verb{node}}{[\verb{\link{GtkCTreeNode}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
