\alias{gtkCTreeNodeGetSelectable}
\name{gtkCTreeNodeGetSelectable}
\title{gtkCTreeNodeGetSelectable}
\description{

\strong{WARNING: \code{gtk_ctree_node_get_selectable} is deprecated and should not be used in newly-written code.  }
}
\usage{gtkCTreeNodeGetSelectable(object, node)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
}
\value{[logical] Whether this node can be selected by the user.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
