\alias{gtkCTreeNodeGetText}
\name{gtkCTreeNodeGetText}
\title{gtkCTreeNodeGetText}
\description{

\strong{WARNING: \code{gtk_ctree_node_get_text} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeGetText(object, node, column)}
\arguments{
\item{\verb{object}}{If nonnull, the pointer to the text string is assigned to *\code{text}.}
\item{\verb{node}}{True if the given cell has text in it.}
\item{\verb{column}}{\emph{undocumented
}}
}
\value{
A list containing the following elements:
\item{retval}{[logical] True if the given cell has text in it.}
\item{\verb{text}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
