\alias{gtkCTreeNodeMoveto}
\name{gtkCTreeNodeMoveto}
\title{gtkCTreeNodeMoveto}
\description{
This function makes the given column of the given node visible by
scrolling.
\strong{WARNING: \code{gtk_ctree_node_moveto} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeMoveto(object, node, column, row.align, col.align)}
\arguments{
\item{\verb{object}}{The node to be made visible.}
\item{\verb{node}}{The column to be made visible.}
\item{\verb{column}}{Where in the window the row should appear.}
\item{\verb{row.align}}{Where in the window the column should appear.}
\item{\verb{col.align}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
