\alias{gtkCTreeNodeSetCellStyle}
\name{gtkCTreeNodeSetCellStyle}
\title{gtkCTreeNodeSetCellStyle}
\description{
Set the style of an individual cell.
\strong{WARNING: \code{gtk_ctree_node_set_cell_style} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetCellStyle(object, node, column, style)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCTree}}] }
\item{\verb{node}}{[\verb{\link{GtkCTreeNode}}] }
\item{\verb{column}}{[integer] }
\item{\verb{style}}{[\verb{\link{GtkStyle}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
