\alias{gtkCTreeNodeSetCellStyle}
\name{gtkCTreeNodeSetCellStyle}
\title{gtkCTreeNodeSetCellStyle}
\description{
Set the style of an individual cell.
\strong{WARNING: \code{gtk_ctree_node_set_cell_style} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetCellStyle(object, node, column, style)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{node}}{\emph{undocumented
}}
\item{\verb{column}}{\emph{undocumented
}}
\item{\verb{style}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
