\alias{gtkCTreeNodeSetForeground}
\name{gtkCTreeNodeSetForeground}
\title{gtkCTreeNodeSetForeground}
\description{

\strong{WARNING: \code{gtk_ctree_node_set_foreground} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetForeground(object, node, color)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
\item{\code{color}}{[\code{\link{GdkColor}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
