\alias{gtkCTreeNodeSetForeground}
\name{gtkCTreeNodeSetForeground}
\title{gtkCTreeNodeSetForeground}
\description{

\strong{WARNING: \code{gtk_ctree_node_set_foreground} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetForeground(object, node, color)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{node}}{\emph{undocumented
}}
\item{\verb{color}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
