\alias{gtkCTreeNodeSetPixmap}
\name{gtkCTreeNodeSetPixmap}
\title{gtkCTreeNodeSetPixmap}
\description{
FIXME
\strong{WARNING: \code{gtk_ctree_node_set_pixmap} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetPixmap(object, node, column, pixmap, mask = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCTree}}] }
\item{\verb{node}}{[\verb{\link{GtkCTreeNode}}] }
\item{\verb{column}}{[integer] }
\item{\verb{pixmap}}{[\verb{\link{GdkPixmap}}] }
\item{\verb{mask}}{[\verb{\link{GdkBitmap}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
