\alias{gtkCTreeNodeSetPixmap}
\name{gtkCTreeNodeSetPixmap}
\title{gtkCTreeNodeSetPixmap}
\description{
FIXME
\strong{WARNING: \code{gtk_ctree_node_set_pixmap} is deprecated and should not be used in newly-written code.  
FIXME}
}
\usage{gtkCTreeNodeSetPixmap(object, node, column, pixmap, mask = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
\item{\code{column}}{[integer] }
\item{\code{pixmap}}{[\code{\link{GdkPixmap}}] }
\item{\code{mask}}{[\code{\link{GdkBitmap}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
