\alias{gtkCTreeNodeSetPixtext}
\name{gtkCTreeNodeSetPixtext}
\title{gtkCTreeNodeSetPixtext}
\description{
FIXME
\strong{WARNING: \code{gtk_ctree_node_set_pixtext} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetPixtext(object, node, column, text, spacing, pixmap, mask = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
\item{\code{column}}{[integer] }
\item{\code{text}}{[character] }
\item{\code{spacing}}{[integer] }
\item{\code{pixmap}}{[\code{\link{GdkPixmap}}] }
\item{\code{mask}}{[\code{\link{GdkBitmap}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
