\alias{gtkCTreeNodeSetPixtext}
\name{gtkCTreeNodeSetPixtext}
\title{gtkCTreeNodeSetPixtext}
\description{
FIXME
\strong{WARNING: \code{gtk_ctree_node_set_pixtext} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetPixtext(object, node, column, text, spacing, pixmap, 
    mask = NULL)}
\arguments{
\item{\verb{object}}{. \emph{[  \acronym{allow-none}  ]}}
\item{\verb{node}}{\emph{undocumented
}}
\item{\verb{column}}{\emph{undocumented
}}
\item{\verb{text}}{\emph{undocumented
}}
\item{\verb{spacing}}{\emph{undocumented
}}
\item{\verb{pixmap}}{\emph{undocumented
}}
\item{\verb{mask}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
