\alias{gtkCTreeNodeSetRowData}
\name{gtkCTreeNodeSetRowData}
\title{gtkCTreeNodeSetRowData}
\description{
Set the custom data associated with a node.
\strong{WARNING: \code{gtk_ctree_node_set_row_data} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetRowData(object, node, data)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
\item{\code{data}}{[R object] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
