\alias{gtkCTreeNodeSetRowData}
\name{gtkCTreeNodeSetRowData}
\title{gtkCTreeNodeSetRowData}
\description{
Set the custom data associated with a node.
\strong{WARNING: \code{gtk_ctree_node_set_row_data} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetRowData(object, node, data)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{node}}{\emph{undocumented
}}
\item{\verb{data}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
