\alias{gtkCTreeNodeSetRowDataFull}
\name{gtkCTreeNodeSetRowDataFull}
\title{gtkCTreeNodeSetRowDataFull}
\description{
This is the full interface to setting row data, so that a destructor
can be given for the data.
\strong{WARNING: \code{gtk_ctree_node_set_row_data_full} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetRowDataFull(object, node, data)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCTree}}] }
\item{\verb{node}}{[\verb{\link{GtkCTreeNode}}] }
\item{\verb{data}}{[R object] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
