\alias{gtkCTreeNodeSetRowDataFull}
\name{gtkCTreeNodeSetRowDataFull}
\title{gtkCTreeNodeSetRowDataFull}
\description{
This is the full interface to setting row data, so that a destructor
can be given for the data.
\strong{WARNING: \code{gtk_ctree_node_set_row_data_full} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetRowDataFull(object, node, data)}
\arguments{
\item{\verb{object}}{The routine to be called when \code{data} is no longer needed.}
\item{\verb{node}}{\emph{undocumented
}}
\item{\verb{data}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
