\alias{gtkCTreeNodeSetRowStyle}
\name{gtkCTreeNodeSetRowStyle}
\title{gtkCTreeNodeSetRowStyle}
\description{
Set the style of a row.
\strong{WARNING: \code{gtk_ctree_node_set_row_style} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetRowStyle(object, node, style)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCTree}}] }
\item{\verb{node}}{[\verb{\link{GtkCTreeNode}}] }
\item{\verb{style}}{[\verb{\link{GtkStyle}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
