\alias{gtkCTreeNodeSetSelectable}
\name{gtkCTreeNodeSetSelectable}
\title{gtkCTreeNodeSetSelectable}
\description{

\strong{WARNING: \code{gtk_ctree_node_set_selectable} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetSelectable(object, node, selectable)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
\item{\code{selectable}}{[logical] Whether this node can be selected by the user.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
