\alias{gtkCTreeNodeSetSelectable}
\name{gtkCTreeNodeSetSelectable}
\title{gtkCTreeNodeSetSelectable}
\description{

\strong{WARNING: \code{gtk_ctree_node_set_selectable} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetSelectable(object, node, selectable)}
\arguments{
\item{\verb{object}}{Whether this node can be selected by the user.}
\item{\verb{node}}{\emph{undocumented
}}
\item{\verb{selectable}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
