\alias{gtkCTreeNodeSetShift}
\name{gtkCTreeNodeSetShift}
\title{gtkCTreeNodeSetShift}
\description{
Shift the given cell the given amounts in pixels.
\strong{WARNING: \code{gtk_ctree_node_set_shift} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetShift(object, node, column, vertical, horizontal)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{node}}{\emph{undocumented
}}
\item{\verb{column}}{\emph{undocumented
}}
\item{\verb{vertical}}{\emph{undocumented
}}
\item{\verb{horizontal}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
