\alias{gtkCTreeNodeSetText}
\name{gtkCTreeNodeSetText}
\title{gtkCTreeNodeSetText}
\description{
Set the text in a node.
\strong{WARNING: \code{gtk_ctree_node_set_text} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeNodeSetText(object, node, column, text)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{node}}{\emph{undocumented
}}
\item{\verb{column}}{The column whose text to change.}
\item{\verb{text}}{The new text.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
