\alias{gtkCTreeNodeSetText}
\name{gtkCTreeNodeSetText}
\title{gtkCTreeNodeSetText}
\description{
Set the text in a node.
\strong{WARNING: \code{gtk_ctree_node_set_text} is deprecated and should not be used in newly-written code.  
Set the text in a node.}
}
\usage{gtkCTreeNodeSetText(object, node, column, text)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
\item{\code{column}}{[integer] The column whose text to change.}
\item{\code{text}}{[character] The new text.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
