\alias{gtkCTreeRealSelectRecursive}
\name{gtkCTreeRealSelectRecursive}
\title{gtkCTreeRealSelectRecursive}
\description{
The function that implements both \code{\link{gtkCTreeSelectRecursive}} and
\code{\link{gtkCTreeUnselectRecursive}}.
\strong{WARNING: \code{gtk_ctree_real_select_recursive} is deprecated and should not be used in newly-written code.  
The function that implements both \code{\link{gtkCTreeSelectRecursive}} and
\code{\link{gtkCTreeUnselectRecursive}}.}
}
\usage{gtkCTreeRealSelectRecursive(object, node, state)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
\item{\code{state}}{[integer] True for selecting, false for unselecting.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
