\alias{gtkCTreeSelect}
\name{gtkCTreeSelect}
\title{gtkCTreeSelect}
\description{
Cause the given node to be selected and emit the appropriate signal.
\strong{WARNING: \code{gtk_ctree_select} is deprecated and should not be used in newly-written code.  
Cause the given node to be selected and emit the appropriate signal.}
}
\usage{gtkCTreeSelect(object, node)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
