\alias{gtkCTreeSelect}
\name{gtkCTreeSelect}
\title{gtkCTreeSelect}
\description{
Cause the given node to be selected and emit the appropriate signal.
\strong{WARNING: \code{gtk_ctree_select} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeSelect(object, node)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{node}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
