\alias{gtkCTreeSelectRecursive}
\name{gtkCTreeSelectRecursive}
\title{gtkCTreeSelectRecursive}
\description{
Cause the given node and its subnodes to be selected and emit the appropriate signal(s).
\strong{WARNING: \code{gtk_ctree_select_recursive} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeSelectRecursive(object, node)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{node}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
