\alias{gtkCTreeSetExpanderStyle}
\name{gtkCTreeSetExpanderStyle}
\title{gtkCTreeSetExpanderStyle}
\description{

\strong{WARNING: \code{gtk_ctree_set_expander_style} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeSetExpanderStyle(object, expander.style)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{expander.style}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
