\alias{gtkCTreeSetLineStyle}
\name{gtkCTreeSetLineStyle}
\title{gtkCTreeSetLineStyle}
\description{

\strong{WARNING: \code{gtk_ctree_set_line_style} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeSetLineStyle(object, line.style)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{line.style}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
