\alias{gtkCTreeSetNodeInfo}
\name{gtkCTreeSetNodeInfo}
\title{gtkCTreeSetNodeInfo}
\description{
Change the information. Most parameters correspond
to the parameters of \code{\link{gtkCTreeInsertNode}}.
\strong{WARNING: \code{gtk_ctree_set_node_info} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeSetNodeInfo(object, node, text, spacing, pixmap.closed = NULL, mask.closed = NULL, pixmap.opened = NULL, mask.opened = NULL, is.leaf, expanded)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{node}}{\emph{undocumented
}}
\item{\verb{text}}{The text to be in the tree column.}
\item{\verb{spacing}}{\emph{undocumented
}}
\item{\verb{pixmap.closed}}{. \emph{[  \acronym{allow-none}  ]}}
\item{\verb{mask.closed}}{. \emph{[  \acronym{allow-none}  ]}}
\item{\verb{pixmap.opened}}{. \emph{[  \acronym{allow-none}  ]}}
\item{\verb{mask.opened}}{. \emph{[  \acronym{allow-none}  ]}}
\item{\verb{is.leaf}}{\emph{undocumented
}}
\item{\verb{expanded}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
