\alias{gtkCTreeSetShowStub}
\name{gtkCTreeSetShowStub}
\title{gtkCTreeSetShowStub}
\description{

\strong{WARNING: \code{gtk_ctree_set_show_stub} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeSetShowStub(object, show.stub)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{show.stub}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
