\alias{gtkCTreeSetSpacing}
\name{gtkCTreeSetSpacing}
\title{gtkCTreeSetSpacing}
\description{
The spacing between the tree graphic and the actual node content.
\strong{WARNING: \code{gtk_ctree_set_spacing} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeSetSpacing(object, spacing)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{spacing}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
