\alias{gtkCTreeSortNode}
\name{gtkCTreeSortNode}
\title{gtkCTreeSortNode}
\description{
Sort the children of a node. See \code{\link{GtkCList}} for how to set the sorting
criteria etc.
\strong{WARNING: \code{gtk_ctree_sort_node} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeSortNode(object, node)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{node}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
