\alias{gtkCTreeSortRecursive}
\name{gtkCTreeSortRecursive}
\title{gtkCTreeSortRecursive}
\description{
Sort the descendants of a node. See \code{\link{GtkCList}} for how to set the sorting
criteria etc.
\strong{WARNING: \code{gtk_ctree_sort_recursive} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeSortRecursive(object, node)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
