\alias{gtkCTreeToggleExpansion}
\name{gtkCTreeToggleExpansion}
\title{gtkCTreeToggleExpansion}
\description{
Toggle a node, i.e. if it is collapsed, expand it and vice versa.
\strong{WARNING: \code{gtk_ctree_toggle_expansion} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeToggleExpansion(object, node)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{node}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
