\alias{gtkCTreeToggleExpansionRecursive}
\name{gtkCTreeToggleExpansionRecursive}
\title{gtkCTreeToggleExpansionRecursive}
\description{
Toggle the expansion of a node and all its children.
\strong{WARNING: \code{gtk_ctree_toggle_expansion_recursive} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeToggleExpansionRecursive(object, node)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCTree}}] }
\item{\verb{node}}{[\verb{\link{GtkCTreeNode}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
