\alias{gtkCTreeUnselect}
\name{gtkCTreeUnselect}
\title{gtkCTreeUnselect}
\description{
Unselect the given node and emit the appropriate signal.
\strong{WARNING: \code{gtk_ctree_unselect} is deprecated and should not be used in newly-written code.  
Unselect the given node and emit the appropriate signal.}
}
\usage{gtkCTreeUnselect(object, node)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
