\alias{gtkCTreeUnselectRecursive}
\name{gtkCTreeUnselectRecursive}
\title{gtkCTreeUnselectRecursive}
\description{
Unselect the given node and its subnodes and emit the appropriate signal(s).
\strong{WARNING: \code{gtk_ctree_unselect_recursive} is deprecated and should not be used in newly-written code.  
Unselect the given node and its subnodes and emit the appropriate signal(s).}
}
\usage{gtkCTreeUnselectRecursive(object, node)}
\arguments{
\item{\code{object}}{[\code{\link{GtkCTree}}] }
\item{\code{node}}{[\code{\link{GtkCTreeNode}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
