\alias{gtkCTreeUnselectRecursive}
\name{gtkCTreeUnselectRecursive}
\title{gtkCTreeUnselectRecursive}
\description{
Unselect the given node and its subnodes and emit the appropriate signal(s).
\strong{WARNING: \code{gtk_ctree_unselect_recursive} is deprecated and should not be used in newly-written code.}
}
\usage{gtkCTreeUnselectRecursive(object, node)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{node}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
