\alias{gtkCurveGetVector}
\name{gtkCurveGetVector}
\title{gtkCurveGetVector}
\description{Returns a vector of points representing the curve.}
\usage{gtkCurveGetVector(object, veclen)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkCurve}}] a \verb{\link{GtkCurve}}.}
\item{\verb{veclen}}{[integer] the number of points to calculate.}
}
\value{
A list containing the following elements:
\item{\verb{vector}}{[numeric] returns the points.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
