\alias{gtkCurveNew}
\name{gtkCurveNew}
\title{gtkCurveNew}
\description{
Creates a new \code{\link{GtkCurve}}.
\strong{WARNING: \code{gtk_curve_new} has been deprecated since version 2.20 and should not be used in newly-written code. Don't use this widget anymore.}
}
\usage{gtkCurveNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkCurve}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
