\alias{gtkCurveReset}
\name{gtkCurveReset}
\title{gtkCurveReset}
\description{Resets the curve to a straight line from the minimum x and y values to the
maximum x and y values (i.e. from the bottom-left to the top-right corners).
The curve type is not changed.}
\usage{gtkCurveReset(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkCurve}}] a \code{\link{GtkCurve}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
