\alias{gtkCurveReset}
\name{gtkCurveReset}
\title{gtkCurveReset}
\description{
Resets the curve to a straight line from the minimum x and y values to the
maximum x and y values (i.e. from the bottom-left to the top-right corners).
The curve type is not changed.
\strong{WARNING: \code{gtk_curve_reset} has been deprecated since version 2.20 and should not be used in newly-written code. Don't use this widget anymore.}
}
\usage{gtkCurveReset(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkCurve}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
