\alias{gtkCurveSetVector}
\name{gtkCurveSetVector}
\title{gtkCurveSetVector}
\description{
Sets the vector of points on the curve.
The curve type is set to \code{GTK_CURVE_TYPE_FREE}.
\strong{WARNING: \code{gtk_curve_set_vector} has been deprecated since version 2.20 and should not be used in newly-written code. Don't use this widget anymore.}
}
\usage{gtkCurveSetVector(object, vector)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkCurve}}.}
\item{\verb{vector}}{the points on the curve.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
