\alias{gtkDialogNew}
\name{gtkDialogNew}
\title{gtkDialogNew}
\description{Creates a new dialog box. Widgets should not be packed into this \code{\link{GtkWindow}}
directly, but into the \code{vbox} and \code{action.area}, as described above.}
\usage{gtkDialogNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkDialog}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
