\alias{gtkDialogSetDefaultResponse}
\name{gtkDialogSetDefaultResponse}
\title{gtkDialogSetDefaultResponse}
\description{Sets the last widget in the dialog's action area with the given \code{response.id}
as the default widget for the dialog. Pressing "Enter" normally activates
the default widget.}
\usage{gtkDialogSetDefaultResponse(object, response.id)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkDialog}}}
\item{\verb{response.id}}{a response ID}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
