\alias{gtkDialogSetHasSeparator}
\name{gtkDialogSetHasSeparator}
\title{gtkDialogSetHasSeparator}
\description{Sets whether the dialog has a separator above the buttons.
\code{TRUE} by default.}
\usage{gtkDialogSetHasSeparator(object, setting)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkDialog}}]  a \verb{\link{GtkDialog}}}
\item{\verb{setting}}{[logical] \code{TRUE} to have a separator}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
