\alias{gtkDragDestGetTargetList}
\name{gtkDragDestGetTargetList}
\title{gtkDragDestGetTargetList}
\description{Returns the list of targets this widget can accept from
drag-and-drop.}
\usage{gtkDragDestGetTargetList(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\value{[\code{\link{GtkTargetList}}] the \code{\link{GtkTargetList}}, or \code{NULL} if none}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
