\alias{gtkDragSourceGetTargetList}
\name{gtkDragSourceGetTargetList}
\title{gtkDragSourceGetTargetList}
\description{Gets the list of targets this widget can provide for
drag-and-drop.}
\usage{gtkDragSourceGetTargetList(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.4}
\value{[\code{\link{GtkTargetList}}] the \code{\link{GtkTargetList}}, or \code{NULL} if none}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
