\alias{gtkDragSourceSetIconName}
\name{gtkDragSourceSetIconName}
\title{gtkDragSourceSetIconName}
\description{Sets the icon that will be used for drags from a particular source
to a themed icon. See the docs for \code{\link{GtkIconTheme}} for more details.}
\usage{gtkDragSourceSetIconName(widget, icon.name)}
\arguments{
\item{\verb{widget}}{a \code{\link{GtkWidget}}}
\item{\verb{icon.name}}{name of icon to use}
}
\details{Since 2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
