\alias{gtkDragSourceSetIconPixbuf}
\name{gtkDragSourceSetIconPixbuf}
\title{gtkDragSourceSetIconPixbuf}
\description{Sets the icon that will be used for drags from a particular widget
from a \verb{\link{GdkPixbuf}}. GTK+ retains a reference for \code{pixbuf} and will 
release it when it is no longer needed.}
\usage{gtkDragSourceSetIconPixbuf(object, pixbuf)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}}
\item{\verb{pixbuf}}{[\verb{\link{GdkPixbuf}}]  the \verb{\link{GdkPixbuf}} for the drag icon}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
