\alias{gtkDragSourceSetIconPixbuf}
\name{gtkDragSourceSetIconPixbuf}
\title{gtkDragSourceSetIconPixbuf}
\description{Sets the icon that will be used for drags from a particular widget
from a \code{\link{GdkPixbuf}}. GTK+ retains a reference for \code{pixbuf} and will 
release it when it is no longer needed.}
\usage{gtkDragSourceSetIconPixbuf(object, pixbuf)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWidget}}}
\item{\verb{pixbuf}}{the \code{\link{GdkPixbuf}} for the drag icon}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
