\alias{gtkDragSourceSetIconStock}
\name{gtkDragSourceSetIconStock}
\title{gtkDragSourceSetIconStock}
\description{Sets the icon that will be used for drags from a particular source
to a stock icon.}
\usage{gtkDragSourceSetIconStock(object, stock.id)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWidget}}}
\item{\verb{stock.id}}{the ID of the stock icon to use}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
