\alias{gtkDragSourceUnset}
\name{gtkDragSourceUnset}
\title{gtkDragSourceUnset}
\description{Undoes the effects of \code{\link{gtkDragSourceSet}}.}
\usage{gtkDragSourceUnset(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
