\alias{gtkDrawLayout}
\name{gtkDrawLayout}
\title{gtkDrawLayout}
\description{

\strong{WARNING: \code{gtk_draw_layout} is deprecated and should not be used in newly-written code.}
}
\usage{gtkDrawLayout(object, window, state.type, use.text, x, y, layout)}
\arguments{
\item{\code{object}}{[\code{\link{GtkStyle}}] }
\item{\code{window}}{[\code{\link{GdkWindow}}] }
\item{\code{state.type}}{[\code{\link{GtkStateType}}] }
\item{\code{use.text}}{[logical] }
\item{\code{x}}{[integer] }
\item{\code{y}}{[integer] }
\item{\code{layout}}{[\code{\link{PangoLayout}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
