\alias{gtkDrawSlider}
\name{gtkDrawSlider}
\title{gtkDrawSlider}
\description{

\strong{WARNING: \code{gtk_draw_slider} is deprecated and should not be used in newly-written code.  }
}
\usage{gtkDrawSlider(object, window, state.type, shadow.type, x, y, width, height, orientation)}
\arguments{
\item{\code{object}}{[\code{\link{GtkStyle}}] }
\item{\code{window}}{[\code{\link{GdkWindow}}] }
\item{\code{state.type}}{[\code{\link{GtkStateType}}] }
\item{\code{shadow.type}}{[\code{\link{GtkShadowType}}] }
\item{\code{x}}{[integer] }
\item{\code{y}}{[integer] }
\item{\code{width}}{[integer] }
\item{\code{height}}{[integer] }
\item{\code{orientation}}{[\code{\link{GtkOrientation}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
