\alias{gtkEditableCopyClipboard}
\name{gtkEditableCopyClipboard}
\title{gtkEditableCopyClipboard}
\description{Copies the contents of the currently selected content in the editable and
puts it on the clipboard.}
\usage{gtkEditableCopyClipboard(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEditable}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
