\alias{gtkEditableCutClipboard}
\name{gtkEditableCutClipboard}
\title{gtkEditableCutClipboard}
\description{Causes the characters in the current selection to
be copied to the clipboard and then deleted from
the widget.}
\usage{gtkEditableCutClipboard(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkEditable}}] a \verb{\link{GtkEditable}} widget.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
