\alias{gtkEditableCutClipboard}
\name{gtkEditableCutClipboard}
\title{gtkEditableCutClipboard}
\description{Removes the contents of the currently selected content in the editable and
puts it on the clipboard.}
\usage{gtkEditableCutClipboard(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEditable}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
