\alias{gtkEditableDeleteSelection}
\name{gtkEditableDeleteSelection}
\title{gtkEditableDeleteSelection}
\description{Deletes the current contents of the widgets selection and
disclaims the selection.}
\usage{gtkEditableDeleteSelection(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkEditable}}] a \code{\link{GtkEditable}} widget.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
