\alias{gtkEditableDeleteSelection}
\name{gtkEditableDeleteSelection}
\title{gtkEditableDeleteSelection}
\description{Deletes the currently selected text of the editable.
This call doesn't do anything if there is no selected text.}
\usage{gtkEditableDeleteSelection(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEditable}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
