\alias{gtkEditableGetPosition}
\name{gtkEditableGetPosition}
\title{gtkEditableGetPosition}
\description{Retrieves the current position of the cursor relative to the start
of the content of the editable. }
\usage{gtkEditableGetPosition(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEditable}}}}
\details{Note that this position is in characters, not in bytes.}
\value{[integer] the cursor position}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
