\alias{gtkEditablePasteClipboard}
\name{gtkEditablePasteClipboard}
\title{gtkEditablePasteClipboard}
\description{Pastes the content of the clipboard to the current position of the
cursor in the editable.}
\usage{gtkEditablePasteClipboard(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEditable}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
