\alias{gtkEditableSetEditable}
\name{gtkEditableSetEditable}
\title{gtkEditableSetEditable}
\description{Determines if the user can edit the text in the editable
widget or not.}
\usage{gtkEditableSetEditable(object, is.editable)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkEditable}}] a \verb{\link{GtkEditable}} widget.}
\item{\verb{is.editable}}{[logical] \code{TRUE} if the user is allowed to edit the text
  in the widget.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
