\alias{gtkEntryAppendText}
\name{gtkEntryAppendText}
\title{gtkEntryAppendText}
\description{
Appends the given text to the contents of the widget.
\strong{WARNING: \code{gtk_entry_append_text} is deprecated and should not be used in newly-written code.}
}
\usage{gtkEntryAppendText(object, text)}
\arguments{
\item{\code{object}}{[\code{\link{GtkEntry}}] a \code{\link{GtkEntry}}.}
\item{\code{text}}{[character] the text to append.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
