\alias{gtkEntryAppendText}
\name{gtkEntryAppendText}
\title{gtkEntryAppendText}
\description{
Appends the given text to the contents of the widget.
\strong{WARNING: \code{gtk_entry_append_text} has been deprecated since version 2.0 and should not be used in newly-written code. Use \code{\link{gtkEditableInsertText}} instead.}
}
\usage{gtkEntryAppendText(object, text)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkEntry}}}
\item{\verb{text}}{the text to append}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
