\alias{gtkEntryCompletionDeleteAction}
\name{gtkEntryCompletionDeleteAction}
\title{gtkEntryCompletionDeleteAction}
\description{Deletes the action at \code{index.} from \code{completion}'s action list.}
\usage{gtkEntryCompletionDeleteAction(object, index)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkEntryCompletion}}.}
\item{\verb{index}}{The index of the item to Delete.}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
