\alias{gtkEntryCompletionGetMinimumKeyLength}
\name{gtkEntryCompletionGetMinimumKeyLength}
\title{gtkEntryCompletionGetMinimumKeyLength}
\description{Returns the minimum key length as set for \code{completion}.}
\usage{gtkEntryCompletionGetMinimumKeyLength(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkEntryCompletion}}.}}
\details{Since 2.4}
\value{[integer] The currently used minimum key length.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
