\alias{gtkEntryCompletionGetModel}
\name{gtkEntryCompletionGetModel}
\title{gtkEntryCompletionGetModel}
\description{Returns the model the \code{\link{GtkEntryCompletion}} is using as data source.
Returns \code{NULL} if the model is unset.}
\usage{gtkEntryCompletionGetModel(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkEntryCompletion}}]  A \code{\link{GtkEntryCompletion}}.}}
\details{  Since  2.4}
\value{[\code{\link{GtkTreeModel}}]  A \code{\link{GtkTreeModel}}, or \code{NULL} if none is currently being used.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
