\alias{gtkEntryCompletionGetModel}
\name{gtkEntryCompletionGetModel}
\title{gtkEntryCompletionGetModel}
\description{Returns \code{NULL} if the model is unset.}
\usage{gtkEntryCompletionGetModel(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkEntryCompletion}}.}}
\details{Since 2.4}
\value{[\code{\link{GtkTreeModel}}] A \code{\link{GtkTreeModel}}, or \code{NULL} if none is currently being used.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
