\alias{gtkEntryCompletionGetTextColumn}
\name{gtkEntryCompletionGetTextColumn}
\title{gtkEntryCompletionGetTextColumn}
\description{Returns the column in the model of \code{completion} to get strings from.}
\usage{gtkEntryCompletionGetTextColumn(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEntryCompletion}}}}
\details{Since 2.6}
\value{[integer] the column containing the strings}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
