\alias{gtkEntryCompletionInsertActionMarkup}
\name{gtkEntryCompletionInsertActionMarkup}
\title{gtkEntryCompletionInsertActionMarkup}
\description{Inserts an action in \code{completion}'s action item list at position \code{index.}
with markup \code{markup}.}
\usage{gtkEntryCompletionInsertActionMarkup(object, index, markup)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkEntryCompletion}}.}
\item{\verb{index}}{The index of the item to insert.}
\item{\verb{markup}}{Markup of the item to insert.}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
