\alias{gtkEntryCompletionInsertPrefix}
\name{gtkEntryCompletionInsertPrefix}
\title{gtkEntryCompletionInsertPrefix}
\description{Requests a prefix insertion.}
\usage{gtkEntryCompletionInsertPrefix(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkEntryCompletion}}]  a \code{\link{GtkEntryCompletion}}}}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
