\alias{gtkEntryCompletionNew}
\name{gtkEntryCompletionNew}
\title{gtkEntryCompletionNew}
\description{Creates a new \code{\link{GtkEntryCompletion}} object.}
\usage{gtkEntryCompletionNew()}
\details{Since 2.4}
\value{[\code{\link{GtkEntryCompletion}}] A newly created \code{\link{GtkEntryCompletion}} object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
