\alias{gtkEntryCompletionSetInlineCompletion}
\name{gtkEntryCompletionSetInlineCompletion}
\title{gtkEntryCompletionSetInlineCompletion}
\description{Sets whether the common prefix of the possible completions should
be automatically inserted in the entry.}
\usage{gtkEntryCompletionSetInlineCompletion(object, inline.completion)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkEntryCompletion}}}
\item{\verb{inline.completion}}{\code{TRUE} to do inline completion}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
