\alias{gtkEntryCompletionSetPopupCompletion}
\name{gtkEntryCompletionSetPopupCompletion}
\title{gtkEntryCompletionSetPopupCompletion}
\description{Sets whether the completions should be presented in a popup window.}
\usage{gtkEntryCompletionSetPopupCompletion(object, popup.completion)}
\arguments{
\item{\code{object}}{[\code{\link{GtkEntryCompletion}}]  a \code{\link{GtkEntryCompletion}}}
\item{\code{popup.completion}}{[logical] \code{TRUE} to do popup completion}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
