\alias{gtkEntryGetActivatesDefault}
\name{gtkEntryGetActivatesDefault}
\title{gtkEntryGetActivatesDefault}
\description{Retrieves the value set by \code{\link{gtkEntrySetActivatesDefault}}.}
\usage{gtkEntryGetActivatesDefault(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkEntry}}]  a \code{\link{GtkEntry}}}}
\value{[logical]  \code{TRUE} if the entry will activate the default widget}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
