\alias{gtkEntryGetCompletion}
\name{gtkEntryGetCompletion}
\title{gtkEntryGetCompletion}
\description{Returns the auxiliary completion object currently in use by \code{entry}.}
\usage{gtkEntryGetCompletion(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkEntry}}}}
\details{Since 2.4}
\value{[\code{\link{GtkEntryCompletion}}] The auxiliary completion object currently in use by \code{entry}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
