\alias{gtkEntryGetCursorHadjustment}
\name{gtkEntryGetCursorHadjustment}
\title{gtkEntryGetCursorHadjustment}
\description{Retrieves the horizontal cursor adjustment for the entry. 
See \code{\link{gtkEntrySetCursorHadjustment}}.}
\usage{gtkEntryGetCursorHadjustment(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEntry}}}}
\details{Since 2.12}
\value{[\code{\link{GtkAdjustment}}] the horizontal cursor adjustment, or \code{NULL}
if none has been set. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
