\alias{gtkEntryGetInnerBorder}
\name{gtkEntryGetInnerBorder}
\title{gtkEntryGetInnerBorder}
\description{This function returns the entry's inner-border property. See
\code{\link{gtkEntrySetInnerBorder}} for more information.}
\usage{gtkEntryGetInnerBorder(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkEntry}}]  a \code{\link{GtkEntry}}}}
\details{  Since  2.10}
\value{[\code{\link{GtkBorder}}]  the entry's \code{\link{GtkBorder}}, or \code{NULL} if none was set.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
