\alias{gtkEntryGetMaxLength}
\name{gtkEntryGetMaxLength}
\title{gtkEntryGetMaxLength}
\description{Retrieves the maximum allowed length of the text in
\code{entry}. See \code{\link{gtkEntrySetMaxLength}}.}
\usage{gtkEntryGetMaxLength(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkEntry}}]  a \code{\link{GtkEntry}}}}
\value{[integer]  the maximum allowed number of characters
              in \code{\link{GtkEntry}}, or 0 if there is no maximum.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
