\alias{gtkEntryGetText}
\name{gtkEntryGetText}
\title{gtkEntryGetText}
\description{Retrieves the contents of the entry widget.
See also \code{\link{gtkEditableGetChars}}.}
\usage{gtkEntryGetText(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkEntry}}]  a \code{\link{GtkEntry}}}}
\value{[character]  a pointer to the contents of the widget as a
     string, modified or
     stored.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
