\alias{gtkEntryGetVisibility}
\name{gtkEntryGetVisibility}
\title{gtkEntryGetVisibility}
\description{Retrieves whether the text in \code{entry} is visible. See
\code{\link{gtkEntrySetVisibility}}.}
\usage{gtkEntryGetVisibility(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkEntry}}}}
\value{[logical] \code{TRUE} if the text is currently visible}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
