\alias{gtkEntryPrependText}
\name{gtkEntryPrependText}
\title{gtkEntryPrependText}
\description{
Prepends the given text to the contents of the widget.
\strong{WARNING: \code{gtk_entry_prepend_text} has been deprecated since version 2.0 and should not be used in newly-written code. Use \code{\link{gtkEditableInsertText}} instead.}
}
\usage{gtkEntryPrependText(object, text)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkEntry}}}
\item{\verb{text}}{the text to prepend}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
