\alias{gtkEntrySetEditable}
\name{gtkEntrySetEditable}
\title{gtkEntrySetEditable}
\description{
Determines if the user can edit the text in the editable
widget or not.
\strong{WARNING: \code{gtk_entry_set_editable} has been deprecated since version 2.0 and should not be used in newly-written code. Use \code{\link{gtkEditableSetEditable}} instead.}
}
\usage{gtkEntrySetEditable(object, editable)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkEntry}}}
\item{\verb{editable}}{\code{TRUE} if the user is allowed to edit the text
in the widget}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
