\alias{gtkEntrySetHasFrame}
\name{gtkEntrySetHasFrame}
\title{gtkEntrySetHasFrame}
\description{Sets whether the entry has a beveled frame around it.}
\usage{gtkEntrySetHasFrame(object, setting)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkEntry}}]  a \verb{\link{GtkEntry}}}
\item{\verb{setting}}{[logical]  new value}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
