\alias{gtkEntrySetHasFrame}
\name{gtkEntrySetHasFrame}
\title{gtkEntrySetHasFrame}
\description{Sets whether the entry has a beveled frame around it.}
\usage{gtkEntrySetHasFrame(object, setting)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkEntry}}}
\item{\verb{setting}}{new value}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
