\alias{gtkEntrySetText}
\name{gtkEntrySetText}
\title{gtkEntrySetText}
\description{Sets the text in the widget to the given
value, replacing the current contents.}
\usage{gtkEntrySetText(object, text)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkEntry}}}
\item{\verb{text}}{the new text}
}
\details{See \code{\link{gtkEntryBufferSetText}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
