\alias{gtkEventsPending}
\name{gtkEventsPending}
\title{gtkEventsPending}
\description{Checks if any events are pending. This can be used to update the GUI
and invoke timeouts etc. while doing some time intensive computation.}
\usage{gtkEventsPending()}
\value{[logical] \code{TRUE} if any events are pending, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
