\alias{gtkExpanderGetSpacing}
\name{gtkExpanderGetSpacing}
\title{gtkExpanderGetSpacing}
\description{Gets the value set by \code{\link{gtkExpanderSetSpacing}}.}
\usage{gtkExpanderGetSpacing(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkExpander}}}}
\details{Since 2.4}
\value{[integer] spacing between the expander and child.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
