\alias{gtkExpanderNew}
\name{gtkExpanderNew}
\title{gtkExpanderNew}
\description{Creates a new expander using \code{label} as the text of the label.}
\usage{gtkExpanderNew(label = NULL, show = TRUE)}
\arguments{\item{\verb{label}}{the text of the label}}
\details{Since 2.4}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkExpander}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
