\alias{gtkExpanderSetLabel}
\name{gtkExpanderSetLabel}
\title{gtkExpanderSetLabel}
\description{Sets the text of the label of the expander to \code{label}.}
\usage{gtkExpanderSetLabel(object, label = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkExpander}}]  a \code{\link{GtkExpander}}}
\item{\code{label}}{[character]  a string}
}
\details{This will also clear any previously set labels.  

  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
