\alias{gtkExpanderSetLabel}
\name{gtkExpanderSetLabel}
\title{gtkExpanderSetLabel}
\description{Sets the text of the label of the expander to \code{label}.}
\usage{gtkExpanderSetLabel(object, label = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkExpander}}}
\item{\verb{label}}{a string. \emph{[  \acronym{allow-none}  ]}}
}
\details{This will also clear any previously set labels.
  Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
