\alias{gtkExpanderSetLabelWidget}
\name{gtkExpanderSetLabelWidget}
\title{gtkExpanderSetLabelWidget}
\description{Set the label widget for the expander. This is the widget
that will appear embedded alongside the expander arrow.}
\usage{gtkExpanderSetLabelWidget(object, label.widget = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkExpander}}}
\item{\verb{label.widget}}{the new label widget. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
