\alias{gtkExpanderSetSpacing}
\name{gtkExpanderSetSpacing}
\title{gtkExpanderSetSpacing}
\description{Sets the spacing field of \code{expander}, which is the number of pixels to
place between expander and the child.}
\usage{gtkExpanderSetSpacing(object, spacing)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkExpander}}}
\item{\verb{spacing}}{distance between the expander and child in pixels.}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
