\alias{gtkExpanderSetUseMarkup}
\name{gtkExpanderSetUseMarkup}
\title{gtkExpanderSetUseMarkup}
\description{Sets whether the text of the label contains markup in Pango's text markup
language. See \code{\link{gtkLabelSetMarkup}}.}
\usage{gtkExpanderSetUseMarkup(object, use.markup)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkExpander}}}
\item{\verb{use.markup}}{\code{TRUE} if the label's text should be parsed for markup}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
