\alias{gtkExpanderSetUseUnderline}
\name{gtkExpanderSetUseUnderline}
\title{gtkExpanderSetUseUnderline}
\description{If true, an underline in the text of the expander label indicates
the next character should be used for the mnemonic accelerator key.}
\usage{gtkExpanderSetUseUnderline(object, use.underline)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkExpander}}}
\item{\verb{use.underline}}{\code{TRUE} if underlines in the text indicate mnemonics}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
