\alias{gtkFileChooserButtonNew}
\name{gtkFileChooserButtonNew}
\title{gtkFileChooserButtonNew}
\description{Creates a new file-selecting button widget.}
\usage{gtkFileChooserButtonNew(title, action, show = TRUE)}
\arguments{
\item{\verb{title}}{[character]  the title of the browse dialog.}
\item{\verb{action}}{[\code{\link{GtkFileChooserAction}}]  the open mode for the widget.}
}
\details{  Since  2.6}
\value{[\verb{\link{GtkWidget}}]  a new button widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
