\alias{gtkFileChooserButtonNew}
\name{gtkFileChooserButtonNew}
\title{gtkFileChooserButtonNew}
\description{Creates a new file-selecting button widget.}
\usage{gtkFileChooserButtonNew(title, action, show = TRUE)}
\arguments{
\item{\verb{title}}{the title of the browse dialog.}
\item{\verb{action}}{the open mode for the widget.}
}
\details{Since 2.6}
\value{[\code{\link{GtkWidget}}] a new button widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
