\alias{gtkFileChooserButtonSetFocusOnClick}
\name{gtkFileChooserButtonSetFocusOnClick}
\title{gtkFileChooserButtonSetFocusOnClick}
\description{Sets whether the button will grab focus when it is clicked with the mouse.
Making mouse clicks not grab focus is useful in places like toolbars where
you don't want the keyboard focus removed from the main area of the
application.}
\usage{gtkFileChooserButtonSetFocusOnClick(object, focus.on.click)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFileChooserButton}}}
\item{\verb{focus.on.click}}{whether the button grabs focus when clicked with the mouse}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
