\alias{gtkFileChooserButtonSetWidthChars}
\name{gtkFileChooserButtonSetWidthChars}
\title{gtkFileChooserButtonSetWidthChars}
\description{Sets the width (in characters) that \code{button} will use to \code{n.chars}.}
\usage{gtkFileChooserButtonSetWidthChars(object, n.chars)}
\arguments{
\item{\verb{object}}{the button widget to examine.}
\item{\verb{n.chars}}{the new width, in characters.}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
