\alias{gtkFileChooserDialogNew}
\name{gtkFileChooserDialogNew}
\title{gtkFileChooserDialogNew}
\description{Creates a new \code{\link{GtkFileChooserDialog}}.  This function is analogous to
\code{\link{gtkDialogNewWithButtons}}.}
\usage{gtkFileChooserDialogNew(title = NULL, parent = NULL, action, ..., show = TRUE)}
\arguments{\item{\code{...}}{[\code{\link{GtkWindow}}]  Transient parent of the dialog, or \code{NULL}}}
\details{  Since  2.4}
\value{[\code{\link{GtkWidget}}]  a new \code{\link{GtkFileChooserDialog}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
