\alias{gtkFileChooserGetFilter}
\name{gtkFileChooserGetFilter}
\title{gtkFileChooserGetFilter}
\description{Gets the current filter; see \code{\link{gtkFileChooserSetFilter}}.}
\usage{gtkFileChooserGetFilter(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFileChooser}}}}
\details{Since 2.4}
\value{[\code{\link{GtkFileFilter}}] the current filter, or \code{NULL}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
