\alias{gtkFileChooserGetSelectMultiple}
\name{gtkFileChooserGetSelectMultiple}
\title{gtkFileChooserGetSelectMultiple}
\description{Gets whether multiple files can be selected in the file
selector. See \code{\link{gtkFileChooserSetSelectMultiple}}.}
\usage{gtkFileChooserGetSelectMultiple(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFileChooser}}}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if multiple files can be selected.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
