\alias{gtkFileChooserListFilters}
\name{gtkFileChooserListFilters}
\title{gtkFileChooserListFilters}
\description{Lists the current set of user-selectable filters; see
\code{\link{gtkFileChooserAddFilter}}, \code{\link{gtkFileChooserRemoveFilter}}.}
\usage{gtkFileChooserListFilters(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFileChooser}}}}
\details{Since 2.4}
\value{[list] a \verb{list} containing the current set of
user selectable filters. \emph{[  \acronym{element-type}   utf8][  \acronym{transfer container}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
