\alias{gtkFileChooserRemoveFilter}
\name{gtkFileChooserRemoveFilter}
\title{gtkFileChooserRemoveFilter}
\description{Removes \code{filter} from the list of filters that the user can select between.}
\usage{gtkFileChooserRemoveFilter(object, filter)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFileChooser}}}
\item{\verb{filter}}{a \code{\link{GtkFileFilter}}}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
