\alias{gtkFileChooserSetExtraWidget}
\name{gtkFileChooserSetExtraWidget}
\title{gtkFileChooserSetExtraWidget}
\description{Sets an application-supplied widget to provide extra options to the user.}
\usage{gtkFileChooserSetExtraWidget(object, extra.widget)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFileChooser}}}
\item{\verb{extra.widget}}{widget for extra options}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
