\alias{gtkFileChooserSetSelectMultiple}
\name{gtkFileChooserSetSelectMultiple}
\title{gtkFileChooserSetSelectMultiple}
\description{Sets whether multiple files can be selected in the file selector.  This is
only relevant if the action is set to be \code{GTK_FILE_CHOOSER_ACTION_OPEN} or
\code{GTK_FILE_CHOOSER_ACTION_SELECT_FOLDER}.}
\usage{gtkFileChooserSetSelectMultiple(object, select.multiple)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFileChooser}}}
\item{\verb{select.multiple}}{\code{TRUE} if multiple files can be selected.}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
