\alias{gtkFileChooserSetShowHidden}
\name{gtkFileChooserSetShowHidden}
\title{gtkFileChooserSetShowHidden}
\description{Sets whether hidden files and folders are displayed in the file selector.}
\usage{gtkFileChooserSetShowHidden(object, show.hidden)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFileChooser}}}
\item{\verb{show.hidden}}{\code{TRUE} if hidden files and folders should be displayed.}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
