\alias{gtkFileChooserUnselectFilename}
\name{gtkFileChooserUnselectFilename}
\title{gtkFileChooserUnselectFilename}
\description{Unselects a currently selected filename. If the filename
is not in the current directory, does not exist, or
is otherwise not currently selected, does nothing.}
\usage{gtkFileChooserUnselectFilename(object, filename)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFileChooser}}}
\item{\verb{filename}}{the filename to unselect}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
