\alias{gtkFileFilterAddMimeType}
\name{gtkFileFilterAddMimeType}
\title{gtkFileFilterAddMimeType}
\description{Adds a rule allowing a given mime type to \code{filter}.}
\usage{gtkFileFilterAddMimeType(object, mime.type)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkFileFilter}}}
\item{\verb{mime.type}}{name of a MIME type}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
