\alias{gtkFileFilterAddPattern}
\name{gtkFileFilterAddPattern}
\title{gtkFileFilterAddPattern}
\description{Adds a rule allowing a shell style glob to a filter.}
\usage{gtkFileFilterAddPattern(object, pattern)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFileFilter}}}
\item{\verb{pattern}}{a shell style glob}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
