\alias{gtkFileFilterAddPixbufFormats}
\name{gtkFileFilterAddPixbufFormats}
\title{gtkFileFilterAddPixbufFormats}
\description{Adds a rule allowing image files in the formats supported
by GdkPixbuf.}
\usage{gtkFileFilterAddPixbufFormats(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFileFilter}}}}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
