\alias{gtkFileSelectionHideFileopButtons}
\name{gtkFileSelectionHideFileopButtons}
\title{gtkFileSelectionHideFileopButtons}
\description{
Hides the file operation buttons that normally appear at the top of the dialog. Useful if you wish to create a custom file selector, based on \verb{\link{GtkFileSelection}}.
\strong{WARNING: \code{gtk_file_selection_hide_fileop_buttons} is deprecated and should not be used in newly-written code.}
}
\usage{gtkFileSelectionHideFileopButtons(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkFileSelection}}] a \verb{\link{GtkFileSelection}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
