\alias{gtkFileSelectionShowFileopButtons}
\name{gtkFileSelectionShowFileopButtons}
\title{gtkFileSelectionShowFileopButtons}
\description{
Shows the file operation buttons, if they have previously been hidden. The rest of the widgets in the dialog will be resized accordingly.
\strong{WARNING: \code{gtk_file_selection_show_fileop_buttons} is deprecated and should not be used in newly-written code.}
}
\usage{gtkFileSelectionShowFileopButtons(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFileSelection}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
