\alias{gtkFixedGetHasWindow}
\name{gtkFixedGetHasWindow}
\title{gtkFixedGetHasWindow}
\description{Gets whether the \code{\link{GtkFixed}} has its own \code{\link{GdkWindow}}.
See \code{gdkFixedSetHasWindow()}.}
\usage{gtkFixedGetHasWindow(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkFixed}}]  a \code{\link{GtkWidget}}}}
\value{[logical]  \code{TRUE} if \code{fixed} has its own window.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
