\alias{gtkFixedMove}
\name{gtkFixedMove}
\title{gtkFixedMove}
\description{Moves a child of a \verb{\link{GtkFixed}} container to the given position.}
\usage{gtkFixedMove(object, widget, x, y)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkFixed}}] a \verb{\link{GtkFixed}}.}
\item{\verb{widget}}{[\verb{\link{GtkWidget}}] the child widget.}
\item{\verb{x}}{[integer] the horizontal position to move the widget to.}
\item{\verb{y}}{[integer] the vertical position to move the widget to.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
