\alias{gtkFixedMove}
\name{gtkFixedMove}
\title{gtkFixedMove}
\description{Moves a child of a \code{\link{GtkFixed}} container to the given position.}
\usage{gtkFixedMove(object, widget, x, y)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFixed}}.}
\item{\verb{widget}}{the child widget.}
\item{\verb{x}}{the horizontal position to move the widget to.}
\item{\verb{y}}{the vertical position to move the widget to.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
