\alias{gtkFixedPut}
\name{gtkFixedPut}
\title{gtkFixedPut}
\description{Adds a widget to a \verb{\link{GtkFixed}} container at the given position.}
\usage{gtkFixedPut(object, widget, x, y)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkFixed}}] a \verb{\link{GtkFixed}}.}
\item{\verb{widget}}{[\verb{\link{GtkWidget}}] the widget to add.}
\item{\verb{x}}{[integer] the horizontal position to place the widget at.}
\item{\verb{y}}{[integer] the vertical position to place the widget at.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
