\alias{gtkFixedPut}
\name{gtkFixedPut}
\title{gtkFixedPut}
\description{Adds a widget to a \code{\link{GtkFixed}} container at the given position.}
\usage{gtkFixedPut(object, widget, x, y)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFixed}}.}
\item{\verb{widget}}{the widget to add.}
\item{\verb{x}}{the horizontal position to place the widget at.}
\item{\verb{y}}{the vertical position to place the widget at.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
