\alias{gtkFontButtonGetFontName}
\name{gtkFontButtonGetFontName}
\title{gtkFontButtonGetFontName}
\description{Retrieves the name of the currently selected font.}
\usage{gtkFontButtonGetFontName(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkFontButton}}]  a \code{\link{GtkFontButton}}}}
\details{  Since  2.4}
\value{[character] .}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
