\alias{gtkFontButtonGetTitle}
\name{gtkFontButtonGetTitle}
\title{gtkFontButtonGetTitle}
\description{Retrieves the title of the font selection dialog.}
\usage{gtkFontButtonGetTitle(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFontButton}}}}
\details{Since 2.4}
\value{[character] an internal copy of the title string which must not be freed.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
