\alias{gtkFontButtonSetTitle}
\name{gtkFontButtonSetTitle}
\title{gtkFontButtonSetTitle}
\description{Sets the title for the font selection dialog.}
\usage{gtkFontButtonSetTitle(object, title)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFontButton}}}
\item{\verb{title}}{a string containing the font selection dialog title}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
