\alias{gtkFontButtonSetUseFont}
\name{gtkFontButtonSetUseFont}
\title{gtkFontButtonSetUseFont}
\description{If \code{use.font} is \code{TRUE}, the font name will be written using the selected font.}
\usage{gtkFontButtonSetUseFont(object, use.font)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFontButton}}}
\item{\verb{use.font}}{If \code{TRUE}, font name will be written using font chosen.}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
