\alias{gtkFontSelectionDialogGetFont}
\name{gtkFontSelectionDialogGetFont}
\title{gtkFontSelectionDialogGetFont}
\description{
Gets the currently-selected font.
\strong{WARNING: \code{gtk_font_selection_dialog_get_font} is deprecated and should not be used in newly-written code.}
}
\usage{gtkFontSelectionDialogGetFont(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkFontSelectionDialog}}] a \code{\link{GtkFontSelectionDialog}}.}}
\value{[\code{\link{GdkFont}}] the currently-selected font, or \code{NULL} if no font is selected.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
