\alias{gtkFontSelectionDialogGetPreviewText}
\name{gtkFontSelectionDialogGetPreviewText}
\title{gtkFontSelectionDialogGetPreviewText}
\description{Gets the text displayed in the preview area.}
\usage{gtkFontSelectionDialogGetPreviewText(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFontSelectionDialog}}}}
\value{[character] the text displayed in the preview area.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
