\alias{gtkFontSelectionDialogNew}
\name{gtkFontSelectionDialogNew}
\title{gtkFontSelectionDialogNew}
\description{Creates a new \code{\link{GtkFontSelectionDialog}}.}
\usage{gtkFontSelectionDialogNew(title = NULL, show = TRUE)}
\arguments{\item{\code{title}}{[character] the title of the dialog box.}}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkFontSelectionDialog}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
