\alias{gtkFontSelectionDialogNew}
\name{gtkFontSelectionDialogNew}
\title{gtkFontSelectionDialogNew}
\description{Creates a new \verb{\link{GtkFontSelectionDialog}}.}
\usage{gtkFontSelectionDialogNew(title = NULL, show = TRUE)}
\arguments{\item{\verb{title}}{[character] the title of the dialog box.}}
\value{[\verb{\link{GtkWidget}}] a new \verb{\link{GtkFontSelectionDialog}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
