\alias{gtkFontSelectionDialogSetFontName}
\name{gtkFontSelectionDialogSetFontName}
\title{gtkFontSelectionDialogSetFontName}
\description{Sets the currently-selected font.}
\usage{gtkFontSelectionDialogSetFontName(object, fontname)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkFontSelectionDialog}}] a \verb{\link{GtkFontSelectionDialog}}.}
\item{\verb{fontname}}{[character] a fontname.}
}
\value{[logical] \code{TRUE} if the font was found.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
