\alias{gtkFontSelectionDialogSetFontName}
\name{gtkFontSelectionDialogSetFontName}
\title{gtkFontSelectionDialogSetFontName}
\description{Sets the currently selected font.}
\usage{gtkFontSelectionDialogSetFontName(object, fontname)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFontSelectionDialog}}}
\item{\verb{fontname}}{a font name like "Helvetica 12" or "Times Bold 18"}
}
\value{[logical] \code{TRUE} if the font selected in \code{fsd} is now the
\code{fontname} specified, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
