\alias{gtkFontSelectionGetFont}
\name{gtkFontSelectionGetFont}
\title{gtkFontSelectionGetFont}
\description{
Gets the currently-selected font.
\strong{WARNING: \code{gtk_font_selection_get_font} is deprecated and should not be used in newly-written code.}
}
\usage{gtkFontSelectionGetFont(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkFontSelection}}] a \verb{\link{GtkFontSelection}}.}}
\value{[\verb{\link{GdkFont}}] the currently-selected font, or NULL if no font is selected.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
