\alias{gtkFontSelectionGetFont}
\name{gtkFontSelectionGetFont}
\title{gtkFontSelectionGetFont}
\description{
Gets the currently-selected font.
\strong{WARNING: \code{gtk_font_selection_get_font} has been deprecated since version 2.0 and should not be used in newly-written code. Use \code{\link{gtkFontSelectionGetFontName}} instead.}
}
\usage{gtkFontSelectionGetFont(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFontSelection}}}}
\value{[\code{\link{GdkFont}}] A \code{\link{GdkFont}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
