\alias{gtkFontSelectionGetFontName}
\name{gtkFontSelectionGetFontName}
\title{gtkFontSelectionGetFontName}
\description{Gets the currently-selected font name.}
\usage{gtkFontSelectionGetFontName(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkFontSelection}}] a \code{\link{GtkFontSelection}}.}}
\value{[character] the name of the currently selected font, or \code{NULL} if
    no font is selected.
A newly allocated string with the currently-se.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
