\alias{gtkFontSelectionGetPreviewText}
\name{gtkFontSelectionGetPreviewText}
\title{gtkFontSelectionGetPreviewText}
\description{Gets the text displayed in the preview area.}
\usage{gtkFontSelectionGetPreviewText(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkFontSelection}}] a \code{\link{GtkFontSelection}}.}}
\value{[character] the text displayed in the preview area.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
