\alias{gtkFontSelectionGetPreviewText}
\name{gtkFontSelectionGetPreviewText}
\title{gtkFontSelectionGetPreviewText}
\description{Gets the text displayed in the preview area.}
\usage{gtkFontSelectionGetPreviewText(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFontSelection}}}}
\value{[character] the text displayed in the preview area.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
