\alias{gtkFontSelectionNew}
\name{gtkFontSelectionNew}
\title{gtkFontSelectionNew}
\description{Creates a new \verb{\link{GtkFontSelection}}.}
\usage{gtkFontSelectionNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}] a new \verb{\link{GtkFontSelection}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
