\alias{gtkFontSelectionNew}
\name{gtkFontSelectionNew}
\title{gtkFontSelectionNew}
\description{Creates a new \code{\link{GtkFontSelection}}.}
\usage{gtkFontSelectionNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a n ew \code{\link{GtkFontSelection}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
