\alias{gtkFontSelectionSetFontName}
\name{gtkFontSelectionSetFontName}
\title{gtkFontSelectionSetFontName}
\description{Sets the currently-selected font.}
\usage{gtkFontSelectionSetFontName(object, fontname)}
\arguments{
\item{\code{object}}{[\code{\link{GtkFontSelection}}] a \code{\link{GtkFontSelection}}.}
\item{\code{fontname}}{[character] a fontname.}
}
\value{[logical] \code{TRUE} if the font was found.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
