\alias{gtkFontSelectionSetPreviewText}
\name{gtkFontSelectionSetPreviewText}
\title{gtkFontSelectionSetPreviewText}
\description{Sets the text displayed in the preview area.}
\usage{gtkFontSelectionSetPreviewText(object, text)}
\arguments{
\item{\code{object}}{[\code{\link{GtkFontSelection}}] a \code{\link{GtkFontSelection}}.}
\item{\code{text}}{[character] the text to display in the preview area.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
