\alias{gtkFontSelectionSetPreviewText}
\name{gtkFontSelectionSetPreviewText}
\title{gtkFontSelectionSetPreviewText}
\description{Sets the text displayed in the preview area.
The \code{text} is used to show how the selected font looks.}
\usage{gtkFontSelectionSetPreviewText(object, text)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFontSelection}}}
\item{\verb{text}}{the text to display in the preview area}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
