\alias{gtkFrameGetLabelAlign}
\name{gtkFrameGetLabelAlign}
\title{gtkFrameGetLabelAlign}
\description{Retrieves the X and Y alignment of the frame's label. See
\code{\link{gtkFrameSetLabelAlign}}.}
\usage{gtkFrameGetLabelAlign(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFrame}}}}
\value{
A list containing the following elements:
\item{\verb{xalign}}{location to store X alignment of frame's label, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
\item{\verb{yalign}}{location to store X alignment of frame's label, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
