\alias{gtkFrameGetLabelWidget}
\name{gtkFrameGetLabelWidget}
\title{gtkFrameGetLabelWidget}
\description{Retrieves the label widget for the frame. See
\code{\link{gtkFrameSetLabelWidget}}.}
\usage{gtkFrameGetLabelWidget(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkFrame}}}}
\value{[\code{\link{GtkWidget}}] the label widget, or \code{NULL} if there is none.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
