\alias{gtkFrameNew}
\name{gtkFrameNew}
\title{gtkFrameNew}
\description{Creates a new \code{\link{GtkFrame}}, with optional label \code{label}.
If \code{label} is \code{NULL}, the label is omitted.}
\usage{gtkFrameNew(label = NULL, show = TRUE)}
\arguments{\item{\verb{label}}{the text to use as the label of the frame}}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkFrame}} widget}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
