\alias{gtkFrameSetLabel}
\name{gtkFrameSetLabel}
\title{gtkFrameSetLabel}
\description{Sets the text of the label. If \code{label} is \code{NULL},
the current label is removed.}
\usage{gtkFrameSetLabel(object, label = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkFrame}}]  a \verb{\link{GtkFrame}}}
\item{\verb{label}}{[character]  the text to use as the label of the frame}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
