\alias{gtkFrameSetShadowType}
\name{gtkFrameSetShadowType}
\title{gtkFrameSetShadowType}
\description{Sets the shadow type for \code{frame}.}
\usage{gtkFrameSetShadowType(object, type)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkFrame}}]  a \verb{\link{GtkFrame}}}
\item{\verb{type}}{[\code{\link{GtkShadowType}}]  the new \code{\link{GtkShadowType}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
