\alias{gtkFrameSetShadowType}
\name{gtkFrameSetShadowType}
\title{gtkFrameSetShadowType}
\description{Sets the shadow type for \code{frame}.}
\usage{gtkFrameSetShadowType(object, type)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkFrame}}}
\item{\verb{type}}{the new \code{\link{GtkShadowType}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
