\alias{gtkGammaCurveNew}
\name{gtkGammaCurveNew}
\title{gtkGammaCurveNew}
\description{Creates a new \code{\link{GtkGammaCurve}}.}
\usage{gtkGammaCurveNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkGammaCurve}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
