\alias{gtkGammaCurveNew}
\name{gtkGammaCurveNew}
\title{gtkGammaCurveNew}
\description{
Creates a new \code{\link{GtkGammaCurve}}.
\strong{WARNING: \code{gtk_gamma_curve_new} has been deprecated since version 2.20 and should not be used in newly-written code. Don't use this widget anymore.}
}
\usage{gtkGammaCurveNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkGammaCurve}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
