\alias{gtkGcRelease}
\name{gtkGcRelease}
\title{gtkGcRelease}
\description{Releases a \verb{\link{GdkGC}} allocated using \code{\link{gtkGcGet}}.}
\usage{gtkGcRelease(gc)}
\arguments{\item{\verb{gc}}{[\verb{\link{GdkGC}}] a \verb{\link{GdkGC}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
