\alias{gtkGrabGetCurrent}
\name{gtkGrabGetCurrent}
\title{gtkGrabGetCurrent}
\description{Queries the current grab of the default window group.}
\usage{gtkGrabGetCurrent()}
\value{[\code{\link{GtkWidget}}] The widget which currently has the grab or \code{NULL} if no grab is active.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
