\alias{gtkGrabRemove}
\name{gtkGrabRemove}
\title{gtkGrabRemove}
\description{Removes the grab from the given widget. You have to pair calls to \code{\link{gtkGrabAdd}}
and \code{\link{gtkGrabRemove}}.}
\usage{gtkGrabRemove(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkWidget}}] The widget which gives up the grab.}}
\note{
This function is deprecated. You should use \code{GDK_PRIORITY_REDRAW} instead.
This function is deprecated. You should use \code{G_PRIORITY_HIGH} instead.
This function is deprecated. You should use \code{G_PRIORITY_DEFAULT_IDLE} instead.
This function is deprecated. You should use \code{G_PRIORITY_LOW} instead.
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
