\alias{gtkHandleBoxSetHandlePosition}
\name{gtkHandleBoxSetHandlePosition}
\title{gtkHandleBoxSetHandlePosition}
\description{Sets the side of the handlebox where the handle is drawn.}
\usage{gtkHandleBoxSetHandlePosition(object, position)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkHandleBox}}}
\item{\verb{position}}{the side of the handlebox where the handle should be drawn.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
