\alias{gtkHandleBoxSetShadowType}
\name{gtkHandleBoxSetShadowType}
\title{gtkHandleBoxSetShadowType}
\description{Sets the type of shadow to be drawn around the border
of the handle box.}
\usage{gtkHandleBoxSetShadowType(object, type)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkHandleBox}}}
\item{\verb{type}}{the shadow type.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
