\alias{gtkHScrollbarNew}
\name{gtkHScrollbarNew}
\title{gtkHScrollbarNew}
\description{Creates a new horizontal scrollbar.}
\usage{gtkHScrollbarNew(adjustment = NULL, show = TRUE)}
\arguments{\item{\verb{adjustment}}{[\verb{\link{GtkAdjustment}}] the \verb{\link{GtkAdjustment}} to use, or \code{NULL} to create a new adjustment.}}
\value{[\verb{\link{GtkWidget}}] the new \verb{\link{GtkHScrollbar}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
