\alias{gtkHScrollbarNew}
\name{gtkHScrollbarNew}
\title{gtkHScrollbarNew}
\description{Creates a new horizontal scrollbar.}
\usage{gtkHScrollbarNew(adjustment = NULL, show = TRUE)}
\arguments{\item{\verb{adjustment}}{the \code{\link{GtkAdjustment}} to use, or \code{NULL} to create a new adjustment. \emph{[  \acronym{allow-none}  ]}}}
\value{[\code{\link{GtkWidget}}] the new \code{\link{GtkHScrollbar}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
