\alias{gtkIconInfoCopy}
\name{gtkIconInfoCopy}
\title{gtkIconInfoCopy}
\description{Make a copy of a \code{\link{GtkIconInfo}}.}
\usage{gtkIconInfoCopy(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconInfo}}}}
\details{Since 2.4}
\value{[\code{\link{GtkIconInfo}}] the new GtkIconInfo}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
