\alias{gtkIconSetCopy}
\name{gtkIconSetCopy}
\title{gtkIconSetCopy}
\description{Copies \code{icon.set} by value.}
\usage{gtkIconSetCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkIconSet}}]  a \verb{\link{GtkIconSet}}}}
\value{[\verb{\link{GtkIconSet}}]  a new \verb{\link{GtkIconSet}} identical to the first.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
