\alias{gtkIconSetCopy}
\name{gtkIconSetCopy}
\title{gtkIconSetCopy}
\description{Copies \code{icon.set} by value.}
\usage{gtkIconSetCopy(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconSet}}}}
\value{[\code{\link{GtkIconSet}}] a new \code{\link{GtkIconSet}} identical to the first.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
