\alias{gtkIconSetGetSizes}
\name{gtkIconSetGetSizes}
\title{gtkIconSetGetSizes}
\description{Obtains a list of icon sizes this icon set can render.}
\usage{gtkIconSetGetSizes(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkIconSet}}]  a \verb{\link{GtkIconSet}}}}
\value{
A list containing the following elements:
\item{\verb{sizes}}{[\code{\link{GtkIconSize}}]  return location for list of sizes}
\item{\verb{n.sizes}}{[integer]  location to store number of elements in returned list}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
