\alias{gtkIconSetGetSizes}
\name{gtkIconSetGetSizes}
\title{gtkIconSetGetSizes}
\description{Obtains a list of icon sizes this icon set can render.}
\usage{gtkIconSetGetSizes(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconSet}}}}
\value{
A list containing the following elements:
\item{\verb{sizes}}{return location
for list of sizes. \emph{[  \acronym{array}   length=n_sizes][  \acronym{out}  ][  \acronym{type}   int]}}
\item{\verb{n.sizes}}{location to store number of elements in returned list}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
