\alias{gtkIconSizeGetName}
\name{gtkIconSizeGetName}
\title{gtkIconSizeGetName}
\description{Gets the canonical name of the given icon size.}
\usage{gtkIconSizeGetName(size)}
\arguments{\item{\verb{size}}{[\code{\link{GtkIconSize}}]  a \code{\link{GtkIconSize}}.}}
\value{[character]  the name of the given icon size.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
