\alias{gtkIconSizeGetName}
\name{gtkIconSizeGetName}
\title{gtkIconSizeGetName}
\description{Gets the canonical name of the given icon size. The returned string
is statically allocated and should not be freed.}
\usage{gtkIconSizeGetName(size)}
\arguments{\item{\verb{size}}{a \code{\link{GtkIconSize}}. \emph{[  \acronym{type}   int]}}}
\value{[character] the name of the given icon size.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
