\alias{gtkIconSourceCopy}
\name{gtkIconSourceCopy}
\title{gtkIconSourceCopy}
\description{Creates a copy of \code{source}; mostly useful for language bindings.}
\usage{gtkIconSourceCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkIconSource}}]  a \verb{\link{GtkIconSource}}}}
\value{[\verb{\link{GtkIconSource}}]  a new \verb{\link{GtkIconSource}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
