\alias{gtkIconSourceCopy}
\name{gtkIconSourceCopy}
\title{gtkIconSourceCopy}
\description{Creates a copy of \code{source}; mostly useful for language bindings.}
\usage{gtkIconSourceCopy(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconSource}}}}
\value{[\code{\link{GtkIconSource}}] a new \code{\link{GtkIconSource}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
