\alias{gtkIconSourceGetDirectionWildcarded}
\name{gtkIconSourceGetDirectionWildcarded}
\title{gtkIconSourceGetDirectionWildcarded}
\description{Gets the value set by \code{\link{gtkIconSourceSetDirectionWildcarded}}.}
\usage{gtkIconSourceGetDirectionWildcarded(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconSource}}}}
\value{[logical] \code{TRUE} if this icon source is a base for any text direction variant}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
