\alias{gtkIconSourceGetSize}
\name{gtkIconSourceGetSize}
\title{gtkIconSourceGetSize}
\description{Obtains the icon size this source applies to. The return value
is only useful/meaningful if the icon size is \emph{not} wildcarded.}
\usage{gtkIconSourceGetSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconSource}}}}
\value{[\code{\link{GtkIconSize}}] icon size this source matches. \emph{[  \acronym{type}   int]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
