\alias{gtkIconSourceSetFilename}
\name{gtkIconSourceSetFilename}
\title{gtkIconSourceSetFilename}
\description{Sets the name of an image file to use as a base image when creating
icon variants for \code{\link{GtkIconSet}}. The filename must be absolute.}
\usage{gtkIconSourceSetFilename(object, filename)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkIconSource}}}
\item{\verb{filename}}{image file to use}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
