\alias{gtkIconSourceSetIconName}
\name{gtkIconSourceSetIconName}
\title{gtkIconSourceSetIconName}
\description{Sets the name of an icon to look up in the current icon theme
to use as a base image when creating icon variants for \code{\link{GtkIconSet}}.}
\usage{gtkIconSourceSetIconName(object, icon.name)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkIconSource}}}
\item{\verb{icon.name}}{name of icon to use. \emph{[  \acronym{allow-none}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
