\alias{gtkIconSourceSetPixbuf}
\name{gtkIconSourceSetPixbuf}
\title{gtkIconSourceSetPixbuf}
\description{Sets a pixbuf to use as a base image when creating icon variants
for \code{\link{GtkIconSet}}.}
\usage{gtkIconSourceSetPixbuf(object, pixbuf)}
\arguments{
\item{\code{object}}{[\code{\link{GtkIconSource}}]  a \code{\link{GtkIconSource}}}
\item{\code{pixbuf}}{[\code{\link{GdkPixbuf}}]  pixbuf to use as a source}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
