\alias{gtkIconSourceSetPixbuf}
\name{gtkIconSourceSetPixbuf}
\title{gtkIconSourceSetPixbuf}
\description{Sets a pixbuf to use as a base image when creating icon variants
for \code{\link{GtkIconSet}}.}
\usage{gtkIconSourceSetPixbuf(object, pixbuf)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkIconSource}}}
\item{\verb{pixbuf}}{pixbuf to use as a source}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
