\alias{gtkIconThemeGetExampleIconName}
\name{gtkIconThemeGetExampleIconName}
\title{gtkIconThemeGetExampleIconName}
\description{Gets the name of an icon that is representative of the
current theme (for instance, to use when presenting
a list of themes to the user.)}
\usage{gtkIconThemeGetExampleIconName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconTheme}}}}
\details{Since 2.4}
\value{[char] the name of an example icon or \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
