\alias{gtkIconThemeListContexts}
\name{gtkIconThemeListContexts}
\title{gtkIconThemeListContexts}
\description{Gets the list of contexts available within the current
hierarchy of icon themes}
\usage{gtkIconThemeListContexts(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconTheme}}}}
\details{Since 2.12}
\value{[list] a \verb{list} list holding the names of all the
contexts in the theme. \emph{[  \acronym{element-type}   utf8][  \acronym{transfer full}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
