\alias{gtkIconThemeSetScreen}
\name{gtkIconThemeSetScreen}
\title{gtkIconThemeSetScreen}
\description{Sets the screen for an icon theme; the screen is used
to track the user's currently configured icon theme,
which might be different for different screens.}
\usage{gtkIconThemeSetScreen(object, screen)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkIconTheme}}}
\item{\verb{screen}}{a \code{\link{GdkScreen}}}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
