\alias{gtkIconViewEnableModelDragDest}
\name{gtkIconViewEnableModelDragDest}
\title{gtkIconViewEnableModelDragDest}
\description{Turns \code{icon.view} into a drop destination for automatic DND.}
\usage{gtkIconViewEnableModelDragDest(object, targets, actions)}
\arguments{
\item{\code{object}}{[\code{\link{GtkIconView}}]  a \code{\link{GtkIconView}}}
\item{\code{targets}}{[\code{\link{GtkTargetEntry}}]  the table of targets that the drag will support}
\item{\code{actions}}{[\code{\link{GdkDragAction}}]  the bitmask of possible actions for a drag to this
   widget}
}
\details{  Since  2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
