\alias{gtkIconViewGetColumnSpacing}
\name{gtkIconViewGetColumnSpacing}
\title{gtkIconViewGetColumnSpacing}
\description{Returns the value of the ::column-spacing property.}
\usage{gtkIconViewGetColumnSpacing(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkIconView}}]  a \verb{\link{GtkIconView}}}}
\details{  Since  2.6}
\value{[integer]  the space between columns}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
