\alias{gtkIconViewGetModel}
\name{gtkIconViewGetModel}
\title{gtkIconViewGetModel}
\description{Returns the model the \code{\link{GtkIconView}} is based on.  Returns \code{NULL} if the
model is unset.}
\usage{gtkIconViewGetModel(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconView}}}}
\details{Since 2.6}
\value{[\code{\link{GtkTreeModel}}] A \code{\link{GtkTreeModel}}, or \code{NULL} if none is currently being used.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
