\alias{gtkIconViewGetRowSpacing}
\name{gtkIconViewGetRowSpacing}
\title{gtkIconViewGetRowSpacing}
\description{Returns the value of the ::row-spacing property.}
\usage{gtkIconViewGetRowSpacing(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconView}}}}
\details{Since 2.6}
\value{[integer] the space between rows}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
