\alias{gtkIconViewGetSelectionMode}
\name{gtkIconViewGetSelectionMode}
\title{gtkIconViewGetSelectionMode}
\description{Gets the selection mode of the \code{icon.view}.}
\usage{gtkIconViewGetSelectionMode(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkIconView}}.}}
\details{Since 2.6}
\value{[\code{\link{GtkSelectionMode}}] the current selection mode}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
