\alias{gtkIconViewItemActivated}
\name{gtkIconViewItemActivated}
\title{gtkIconViewItemActivated}
\description{Activates the item determined by \code{path}.}
\usage{gtkIconViewItemActivated(object, path)}
\arguments{
\item{\code{object}}{[\code{\link{GtkIconView}}]  A \code{\link{GtkIconView}}}
\item{\code{path}}{[\code{\link{GtkTreePath}}]  The \code{\link{GtkTreePath}} to be activated}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
