\alias{gtkIconViewItemActivated}
\name{gtkIconViewItemActivated}
\title{gtkIconViewItemActivated}
\description{Activates the item determined by \code{path}.}
\usage{gtkIconViewItemActivated(object, path)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkIconView}}}
\item{\verb{path}}{The \code{\link{GtkTreePath}} to be activated}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
