\alias{gtkIconViewNew}
\name{gtkIconViewNew}
\title{gtkIconViewNew}
\description{Creates a new \code{\link{GtkIconView}} widget}
\usage{gtkIconViewNew(show = TRUE)}
\details{Since 2.6}
\value{[\code{\link{GtkWidget}}] A newly created \code{\link{GtkIconView}} widget}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
