\alias{gtkIconViewNewWithModel}
\name{gtkIconViewNewWithModel}
\title{gtkIconViewNewWithModel}
\description{Creates a new \code{\link{GtkIconView}} widget with the model \code{model}.}
\usage{gtkIconViewNewWithModel(model = NULL, show = TRUE)}
\arguments{\item{\verb{model}}{The model.}}
\details{Since 2.6}
\value{[\code{\link{GtkWidget}}] A newly created \code{\link{GtkIconView}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
