\alias{gtkIconViewSelectAll}
\name{gtkIconViewSelectAll}
\title{gtkIconViewSelectAll}
\description{Selects all the icons. \code{icon.view} must has its selection mode set
to \verb{GTK_SELECTION_MULTIPLE}.}
\usage{gtkIconViewSelectAll(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkIconView}}.}}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
