\alias{gtkIconViewSelectedForeach}
\name{gtkIconViewSelectedForeach}
\title{gtkIconViewSelectedForeach}
\description{Calls a function for each selected icon. Note that the model or
selection cannot be modified from within this function.}
\usage{gtkIconViewSelectedForeach(object, func, data = NULL)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkIconView}}.}
\item{\verb{func}}{The funcion to call for each selected icon.}
\item{\verb{data}}{User data to pass to the function.}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
