\alias{gtkIconViewSelectPath}
\name{gtkIconViewSelectPath}
\title{gtkIconViewSelectPath}
\description{Selects the row at \code{path}.}
\usage{gtkIconViewSelectPath(object, path)}
\arguments{
\item{\code{object}}{[\code{\link{GtkIconView}}]  A \code{\link{GtkIconView}}.}
\item{\code{path}}{[\code{\link{GtkTreePath}}]  The \code{\link{GtkTreePath}} to be selected.}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
