\alias{gtkIconViewSelectPath}
\name{gtkIconViewSelectPath}
\title{gtkIconViewSelectPath}
\description{Selects the row at \code{path}.}
\usage{gtkIconViewSelectPath(object, path)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkIconView}}.}
\item{\verb{path}}{The \code{\link{GtkTreePath}} to be selected.}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
