\alias{gtkIconViewSetColumns}
\name{gtkIconViewSetColumns}
\title{gtkIconViewSetColumns}
\description{Sets the ::columns property which determines in how
many columns the icons are arranged. If \code{columns} is
-1, the number of columns will be chosen automatically 
to fill the available area.}
\usage{gtkIconViewSetColumns(object, columns)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkIconView}}}
\item{\verb{columns}}{the number of columns}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
