\alias{gtkIconViewSetColumnSpacing}
\name{gtkIconViewSetColumnSpacing}
\title{gtkIconViewSetColumnSpacing}
\description{Sets the ::column-spacing property which specifies the space 
which is inserted between the columns of the icon view.}
\usage{gtkIconViewSetColumnSpacing(object, column.spacing)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkIconView}}}
\item{\verb{column.spacing}}{the column spacing}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
