\alias{gtkIconViewSetDragDestItem}
\name{gtkIconViewSetDragDestItem}
\title{gtkIconViewSetDragDestItem}
\description{Sets the item that is highlighted for feedback.}
\usage{gtkIconViewSetDragDestItem(object, path, pos)}
\arguments{
\item{\code{object}}{[\code{\link{GtkIconView}}]  a \code{\link{GtkIconView}}}
\item{\code{path}}{[\code{\link{GtkTreePath}}]  The path of the item to highlight, or \code{NULL}.}
\item{\code{pos}}{[\code{\link{GtkIconViewDropPosition}}]  Specifies where to drop, relative to the item}
}
\details{  Since  2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
