\alias{gtkIconViewSetDragDestItem}
\name{gtkIconViewSetDragDestItem}
\title{gtkIconViewSetDragDestItem}
\description{Sets the item that is highlighted for feedback.}
\usage{gtkIconViewSetDragDestItem(object, path, pos)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkIconView}}}
\item{\verb{path}}{The path of the item to highlight, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
\item{\verb{pos}}{Specifies where to drop, relative to the item}
}
\details{Since 2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
