\alias{gtkIconViewSetMargin}
\name{gtkIconViewSetMargin}
\title{gtkIconViewSetMargin}
\description{Sets the ::margin property which specifies the space 
which is inserted at the top, bottom, left and right 
of the icon view.}
\usage{gtkIconViewSetMargin(object, margin)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkIconView}}}
\item{\verb{margin}}{the margin}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
