\alias{gtkIconViewSetOrientation}
\name{gtkIconViewSetOrientation}
\title{gtkIconViewSetOrientation}
\description{Sets the ::orientation property which determines whether the labels 
are drawn beside the icons instead of below.}
\usage{gtkIconViewSetOrientation(object, orientation)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkIconView}}]  a \verb{\link{GtkIconView}}}
\item{\verb{orientation}}{[\code{\link{GtkOrientation}}]  the relative position of texts and icons }
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
