\alias{gtkIconViewSetPixbufColumn}
\name{gtkIconViewSetPixbufColumn}
\title{gtkIconViewSetPixbufColumn}
\description{Sets the column with pixbufs for \code{icon.view} to be \code{column}. The pixbuf
column must be of type \code{GDK_TYPE_PIXBUF}}
\usage{gtkIconViewSetPixbufColumn(object, column)}
\arguments{
\item{\code{object}}{[\code{\link{GtkIconView}}]  A \code{\link{GtkIconView}}.}
\item{\code{column}}{[integer]  A column in the currently used model.}
}
\details{  Since  2.6 }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
