\alias{gtkIconViewSetPixbufColumn}
\name{gtkIconViewSetPixbufColumn}
\title{gtkIconViewSetPixbufColumn}
\description{Sets the column with pixbufs for \code{icon.view} to be \code{column}. The pixbuf
column must be of type \verb{GDK_TYPE_PIXBUF}}
\usage{gtkIconViewSetPixbufColumn(object, column)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkIconView}}.}
\item{\verb{column}}{A column in the currently used model, or -1 to disable}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
