\alias{gtkIconViewSetSelectionMode}
\name{gtkIconViewSetSelectionMode}
\title{gtkIconViewSetSelectionMode}
\description{Sets the selection mode of the \code{icon.view}.}
\usage{gtkIconViewSetSelectionMode(object, mode)}
\arguments{
\item{\code{object}}{[\code{\link{GtkIconView}}]  A \code{\link{GtkIconView}}.}
\item{\code{mode}}{[\code{\link{GtkSelectionMode}}]  The selection mode}
}
\details{  Since  2.6 }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
