\alias{gtkIconViewSetSelectionMode}
\name{gtkIconViewSetSelectionMode}
\title{gtkIconViewSetSelectionMode}
\description{Sets the selection mode of the \code{icon.view}.}
\usage{gtkIconViewSetSelectionMode(object, mode)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkIconView}}.}
\item{\verb{mode}}{The selection mode}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
