\alias{gtkIconViewSetTextColumn}
\name{gtkIconViewSetTextColumn}
\title{gtkIconViewSetTextColumn}
\description{Sets the column with text for \code{icon.view} to be \code{column}. The text
column must be of type \code{G_TYPE_STRING}.}
\usage{gtkIconViewSetTextColumn(object, column)}
\arguments{
\item{\code{object}}{[\code{\link{GtkIconView}}]  A \code{\link{GtkIconView}}.}
\item{\code{column}}{[integer]  A column in the currently used model, or -1 to display no text}
}
\details{  Since  2.6 }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
