\alias{gtkIconViewSetTextColumn}
\name{gtkIconViewSetTextColumn}
\title{gtkIconViewSetTextColumn}
\description{Sets the column with text for \code{icon.view} to be \code{column}. The text
column must be of type \verb{G_TYPE_STRING}.}
\usage{gtkIconViewSetTextColumn(object, column)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkIconView}}.}
\item{\verb{column}}{A column in the currently used model, or -1 to display no text}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
