\alias{gtkIconViewSetTooltipCell}
\name{gtkIconViewSetTooltipCell}
\title{gtkIconViewSetTooltipCell}
\description{Sets the tip area of \code{tooltip} to the area which \code{cell} occupies in
the item pointed to by \code{path}. See also \code{\link{gtkTooltipsetTipArea}}.}
\usage{gtkIconViewSetTooltipCell(object, tooltip, path, cell)}
\arguments{
\item{\code{object}}{[\code{\link{GtkIconView}}]  a \code{\link{GtkIconView}}}
\item{\code{tooltip}}{[\code{\link{GtkTooltip}}]  a \code{\link{GtkTooltip}}}
\item{\code{path}}{[\code{\link{GtkTreePath}}]  a \code{\link{GtkTreePath}}}
\item{\code{cell}}{[\code{\link{GtkCellRenderer}}]  a \code{\link{GtkCellRenderer}} or \code{NULL}}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
