\alias{gtkIconViewSetTooltipItem}
\name{gtkIconViewSetTooltipItem}
\title{gtkIconViewSetTooltipItem}
\description{Sets the tip area of \code{tooltip} to be the area covered by the item at \code{path}.
See also \code{\link{gtkIconViewSetTooltipColumn}} for a simpler alternative.
See also \code{\link{gtkTooltipSetTipArea}}.}
\usage{gtkIconViewSetTooltipItem(object, tooltip, path)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkIconView}}}
\item{\verb{tooltip}}{a \code{\link{GtkTooltip}}}
\item{\verb{path}}{a \code{\link{GtkTreePath}}}
}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
