\alias{gtkIconViewUnselectAll}
\name{gtkIconViewUnselectAll}
\title{gtkIconViewUnselectAll}
\description{Unselects all the icons.}
\usage{gtkIconViewUnselectAll(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkIconView}}]  A \code{\link{GtkIconView}}.}}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
