\alias{gtkIconViewUnselectAll}
\name{gtkIconViewUnselectAll}
\title{gtkIconViewUnselectAll}
\description{Unselects all the icons.}
\usage{gtkIconViewUnselectAll(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkIconView}}.}}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
