\alias{gtkIconViewUnselectPath}
\name{gtkIconViewUnselectPath}
\title{gtkIconViewUnselectPath}
\description{Unselects the row at \code{path}.}
\usage{gtkIconViewUnselectPath(object, path)}
\arguments{
\item{\code{object}}{[\code{\link{GtkIconView}}]  A \code{\link{GtkIconView}}.}
\item{\code{path}}{[\code{\link{GtkTreePath}}]  The \code{\link{GtkTreePath}} to be unselected.}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
