\alias{gtkIconViewUnselectPath}
\name{gtkIconViewUnselectPath}
\title{gtkIconViewUnselectPath}
\description{Unselects the row at \code{path}.}
\usage{gtkIconViewUnselectPath(object, path)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkIconView}}.}
\item{\verb{path}}{The \code{\link{GtkTreePath}} to be unselected.}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
