\alias{gtkIconViewUnsetModelDragDest}
\name{gtkIconViewUnsetModelDragDest}
\title{gtkIconViewUnsetModelDragDest}
\description{Undoes the effect of \code{\link{gtkIconViewEnableModelDragDest}}. Calling this
method sets \verb{"reorderable"} to \code{FALSE}.}
\usage{gtkIconViewUnsetModelDragDest(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconView}}}}
\details{Since 2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
