\alias{gtkIconViewUnsetModelDragSource}
\name{gtkIconViewUnsetModelDragSource}
\title{gtkIconViewUnsetModelDragSource}
\description{Undoes the effect of \code{\link{gtkIconViewEnableModelDragSource}}. Calling this
method sets \verb{"reorderable"} to \code{FALSE}.}
\usage{gtkIconViewUnsetModelDragSource(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkIconView}}}}
\details{Since 2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
