\alias{gtkImageClear}
\name{gtkImageClear}
\title{gtkImageClear}
\description{Resets the image to be empty.}
\usage{gtkImageClear(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkImage}}]  a \verb{\link{GtkImage}}}}
\details{  Since  2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
