\alias{gtkImageGetPixelSize}
\name{gtkImageGetPixelSize}
\title{gtkImageGetPixelSize}
\description{Gets the pixel size used for named icons.}
\usage{gtkImageGetPixelSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkImage}}}}
\details{Since 2.6}
\value{[integer] the pixel size used for named icons.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
