\alias{gtkImageMenuItemNew}
\name{gtkImageMenuItemNew}
\title{gtkImageMenuItemNew}
\description{Creates a new \code{\link{GtkImageMenuItem}} with an empty label.}
\usage{gtkImageMenuItemNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkImageMenuItem}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
