\alias{gtkImageMenuItemNewWithLabel}
\name{gtkImageMenuItemNewWithLabel}
\title{gtkImageMenuItemNewWithLabel}
\description{Creates a new \verb{\link{GtkImageMenuItem}} containing a label.}
\usage{gtkImageMenuItemNewWithLabel(label, show = TRUE)}
\arguments{\item{\verb{label}}{[character]  the text of the menu item.}}
\value{[\verb{\link{GtkWidget}}]  a new \verb{\link{GtkImageMenuItem}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
