\alias{gtkImageMenuItemNewWithLabel}
\name{gtkImageMenuItemNewWithLabel}
\title{gtkImageMenuItemNewWithLabel}
\description{Creates a new \code{\link{GtkImageMenuItem}} containing a label.}
\usage{gtkImageMenuItemNewWithLabel(label, show = TRUE)}
\arguments{\item{\verb{label}}{the text of the menu item.}}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkImageMenuItem}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
