\alias{gtkImageNew}
\name{gtkImageNew}
\title{gtkImageNew}
\description{Creates a new empty \verb{\link{GtkImage}} widget.}
\usage{gtkImageNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}]  a newly created \verb{\link{GtkImage}} widget. }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
