\alias{gtkImageNew}
\name{gtkImageNew}
\title{gtkImageNew}
\description{Creates a new empty \code{\link{GtkImage}} widget.}
\usage{gtkImageNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a newly created \code{\link{GtkImage}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
