\alias{gtkImageSet}
\name{gtkImageSet}
\title{gtkImageSet}
\description{
Sets the \verb{\link{GtkImage}}.
\strong{WARNING: \code{gtk_image_set} is deprecated and should not be used in newly-written code.}
}
\usage{gtkImageSet(object, val, mask)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkImage}}] a \verb{\link{GtkImage}}}
\item{\verb{val}}{[\verb{\link{GdkImage}}] a \verb{\link{GdkImage}}}
\item{\verb{mask}}{[\verb{\link{GdkBitmap}}] a \verb{\link{GdkBitmap}} that indicates which parts of the image should be transparent.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
