\alias{gtkImageSetFromAnimation}
\name{gtkImageSetFromAnimation}
\title{gtkImageSetFromAnimation}
\description{Causes the \code{\link{GtkImage}} to display the given animation (or display
nothing, if you set the animation to \code{NULL}).}
\usage{gtkImageSetFromAnimation(object, animation)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkImage}}}
\item{\verb{animation}}{the \code{\link{GdkPixbufAnimation}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
