\alias{gtkImageSetFromFile}
\name{gtkImageSetFromFile}
\title{gtkImageSetFromFile}
\description{See \code{\link{gtkImageNewFromFile}} for details.}
\usage{gtkImageSetFromFile(object, filename = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkImage}}]  a \code{\link{GtkImage}}}
\item{\code{filename}}{[character]  a filename or \code{NULL}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
