\alias{gtkImageSetFromFile}
\name{gtkImageSetFromFile}
\title{gtkImageSetFromFile}
\description{See \code{\link{gtkImageNewFromFile}} for details.}
\usage{gtkImageSetFromFile(object, filename = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkImage}}}
\item{\verb{filename}}{ a filename or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
