\alias{gtkImageSetFromIconName}
\name{gtkImageSetFromIconName}
\title{gtkImageSetFromIconName}
\description{See \code{\link{gtkImageNewFromIconName}} for details.}
\usage{gtkImageSetFromIconName(object, icon.name, size)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkImage}}}
\item{\verb{icon.name}}{an icon name}
\item{\verb{size}}{an icon size. \emph{[  \acronym{type}   int]}}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
