\alias{gtkImageSetFromIconSet}
\name{gtkImageSetFromIconSet}
\title{gtkImageSetFromIconSet}
\description{See \code{\link{gtkImageNewFromIconSet}} for details.}
\usage{gtkImageSetFromIconSet(object, icon.set, size)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkImage}}]  a \verb{\link{GtkImage}}}
\item{\verb{icon.set}}{[\verb{\link{GtkIconSet}}]  a \verb{\link{GtkIconSet}}}
\item{\verb{size}}{[\code{\link{GtkIconSize}}]  a stock icon size}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
