\alias{gtkImageSetFromIconSet}
\name{gtkImageSetFromIconSet}
\title{gtkImageSetFromIconSet}
\description{See \code{\link{gtkImageNewFromIconSet}} for details.}
\usage{gtkImageSetFromIconSet(object, icon.set, size)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkImage}}}
\item{\verb{icon.set}}{a \code{\link{GtkIconSet}}}
\item{\verb{size}}{a stock icon size. \emph{[  \acronym{type}   int]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
