\alias{gtkImageSetFromImage}
\name{gtkImageSetFromImage}
\title{gtkImageSetFromImage}
\description{See \code{\link{gtkImageNewFromImage}} for details.}
\usage{gtkImageSetFromImage(object, gdk.image = NULL, mask = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkImage}}]  a \verb{\link{GtkImage}}}
\item{\verb{gdk.image}}{[\verb{\link{GdkImage}}]  a \verb{\link{GdkImage}} or \code{NULL}}
\item{\verb{mask}}{[\verb{\link{GdkBitmap}}]  a \verb{\link{GdkBitmap}} or \code{NULL}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
