\alias{gtkImageSetFromImage}
\name{gtkImageSetFromImage}
\title{gtkImageSetFromImage}
\description{See \code{\link{gtkImageNewFromImage}} for details.}
\usage{gtkImageSetFromImage(object, gdk.image = NULL, mask = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkImage}}}
\item{\verb{gdk.image}}{a \code{\link{GdkImage}} or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
\item{\verb{mask}}{a \code{\link{GdkBitmap}} or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
