\alias{gtkImageSetFromPixbuf}
\name{gtkImageSetFromPixbuf}
\title{gtkImageSetFromPixbuf}
\description{See \code{\link{gtkImageNewFromPixbuf}} for details.}
\usage{gtkImageSetFromPixbuf(object, pixbuf = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkImage}}]  a \code{\link{GtkImage}}}
\item{\code{pixbuf}}{[\code{\link{GdkPixbuf}}]  a \code{\link{GdkPixbuf}} or \code{NULL}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
