\alias{gtkImageSetFromPixbuf}
\name{gtkImageSetFromPixbuf}
\title{gtkImageSetFromPixbuf}
\description{See \code{\link{gtkImageNewFromPixbuf}} for details.}
\usage{gtkImageSetFromPixbuf(object, pixbuf = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkImage}}}
\item{\verb{pixbuf}}{a \code{\link{GdkPixbuf}} or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
