\alias{gtkImageSetFromPixmap}
\name{gtkImageSetFromPixmap}
\title{gtkImageSetFromPixmap}
\description{See \code{\link{gtkImageNewFromPixmap}} for details.}
\usage{gtkImageSetFromPixmap(object, pixmap, mask = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkImage}}]  a \code{\link{GtkImage}}}
\item{\code{pixmap}}{[\code{\link{GdkPixmap}}]  a \code{\link{GdkPixmap}} or \code{NULL}}
\item{\code{mask}}{[\code{\link{GdkBitmap}}]  a \code{\link{GdkBitmap}} or \code{NULL}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
