\alias{gtkImageSetFromPixmap}
\name{gtkImageSetFromPixmap}
\title{gtkImageSetFromPixmap}
\description{See \code{\link{gtkImageNewFromPixmap}} for details.}
\usage{gtkImageSetFromPixmap(object, pixmap, mask = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkImage}}}
\item{\verb{pixmap}}{a \code{\link{GdkPixmap}} or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
\item{\verb{mask}}{a \code{\link{GdkBitmap}} or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
