\alias{gtkImageSetFromStock}
\name{gtkImageSetFromStock}
\title{gtkImageSetFromStock}
\description{See \code{\link{gtkImageNewFromStock}} for details.}
\usage{gtkImageSetFromStock(object, stock.id, size)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkImage}}}
\item{\verb{stock.id}}{a stock icon name}
\item{\verb{size}}{a stock icon size. \emph{[  \acronym{type}   int]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
