\alias{gtkIMContextSetCursorLocation}
\name{gtkIMContextSetCursorLocation}
\title{gtkIMContextSetCursorLocation}
\description{Notify the input method that a change in cursor 
position has been made. The location is relative to the client
window.}
\usage{gtkIMContextSetCursorLocation(object, area)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkIMContext}}]  a \verb{\link{GtkIMContext}}}
\item{\verb{area}}{[\verb{\link{GdkRectangle}}]  new location}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
