\alias{gtkIMContextSetCursorLocation}
\name{gtkIMContextSetCursorLocation}
\title{gtkIMContextSetCursorLocation}
\description{Notify the input method that a change in cursor 
position has been made. The location is relative to the client
window.}
\usage{gtkIMContextSetCursorLocation(object, area)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkIMContext}}}
\item{\verb{area}}{new location}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
