\alias{GtkIMContextSimple}
\alias{gtkIMContextSimple}
\name{GtkIMContextSimple}
\title{GtkIMContextSimple}
\description{An input method context supporting table-based input methods}
\usage{
\link{gtkIMContextSimpleNew}()
\link{gtkIMContextSimpleAddTable}(object, data, max.seq.len, n.seqs)
gtkIMContextSimple()
}
\section{Hierarchy}{\preformatted{  \link{GObject}
   +----\link{GtkIMContext}
         +----GtkIMContextSimple}}
\section{Structures}{\describe{\item{\code{GtkIMContextSimple}}{\emph{undocumented
}}}}
\section{Convenient Construction}{\code{gtkIMContextSimple} is the equivalent of \code{\link{gtkIMContextSimpleNew}}.}
\references{\url{http://developer.gnome.org/doc/API/2.0/gtk/GtkIMContextSimple.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
