\alias{gtkIMContextSimpleNew}
\name{gtkIMContextSimpleNew}
\title{gtkIMContextSimpleNew}
\description{Creates a new \verb{\link{GtkIMContextSimple}}.}
\usage{gtkIMContextSimpleNew()}
\value{[\verb{\link{GtkIMContext}}]  a new \verb{\link{GtkIMContextSimple}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
