\alias{gtkIMContextSimpleNew}
\name{gtkIMContextSimpleNew}
\title{gtkIMContextSimpleNew}
\description{Creates a new \code{\link{GtkIMContextSimple}}.}
\usage{gtkIMContextSimpleNew()}
\value{[\code{\link{GtkIMContext}}] a new \code{\link{GtkIMContextSimple}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
