\alias{GtkIMMulticontext}
\alias{gtkIMMulticontext}
\name{GtkIMMulticontext}
\title{GtkIMMulticontext}
\description{An input method context supporting multiple, loadable input methods}
\usage{
\link{gtkIMMulticontextNew}()
\link{gtkIMMulticontextAppendMenuitems}(object, menushell)
gtkIMMulticontext()
}
\section{Hierarchy}{\preformatted{  \link{GObject}
   +----\link{GtkIMContext}
         +----GtkIMMulticontext}}
\section{Structures}{\describe{\item{\code{GtkIMMulticontext}}{\emph{undocumented
}}}}
\section{Convenient Construction}{\code{gtkIMMulticontext} is the equivalent of \code{\link{gtkIMMulticontextNew}}.}
\references{\url{http://developer.gnome.org/doc/API/2.0/gtk/GtkIMMulticontext.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
