\alias{GtkIMMulticontext}
\alias{gtkIMMulticontext}
\name{GtkIMMulticontext}
\title{GtkIMMulticontext}
\description{An input method context supporting multiple, loadable input methods}
\section{Methods and Functions}{
\code{\link{gtkIMMulticontextNew}()}\cr
\code{\link{gtkIMMulticontextAppendMenuitems}(object, menushell)}\cr
\code{\link{gtkIMMulticontextGetContextId}(object)}\cr
\code{\link{gtkIMMulticontextSetContextId}(object, context.id)}\cr
\code{gtkIMMulticontext()}
}
\section{Hierarchy}{\preformatted{GObject
   +----GtkIMContext
         +----GtkIMMulticontext}}
\section{Structures}{\describe{\item{\verb{GtkIMMulticontext}}{
\emph{undocumented
}

}}}
\section{Convenient Construction}{\code{gtkIMMulticontext} is the equivalent of \code{\link{gtkIMMulticontextNew}}.}
\references{\url{https://developer-old.gnome.org/gtk2/stable/GtkIMMulticontext.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
