\alias{gtkIMMulticontextNew}
\name{gtkIMMulticontextNew}
\title{gtkIMMulticontextNew}
\description{Creates a new \code{\link{GtkIMMulticontext}}.}
\usage{gtkIMMulticontextNew()}
\value{[\code{\link{GtkIMContext}}] a new \code{\link{GtkIMMulticontext}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
