\alias{gtkInitAdd}
\name{gtkInitAdd}
\title{gtkInitAdd}
\description{Registers a function to be called when the mainloop is started.}
\usage{gtkInitAdd(fun, data = NULL)}
\arguments{\item{\verb{data}}{Data to pass to that function.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
