\alias{gtkInputDialogNew}
\name{gtkInputDialogNew}
\title{gtkInputDialogNew}
\description{Creates a new \verb{\link{GtkInputDialog}}.}
\usage{gtkInputDialogNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}] the new \verb{\link{GtkInputDialog}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
