\alias{gtkInputDialogNew}
\name{gtkInputDialogNew}
\title{gtkInputDialogNew}
\description{
Creates a new \code{\link{GtkInputDialog}}.
\strong{WARNING: \code{gtk_input_dialog_new} has been deprecated since version 2.20 and should not be used in newly-written code. Don't use this widget anymore.}
}
\usage{gtkInputDialogNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] the new \code{\link{GtkInputDialog}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
