\alias{gtkInputRemove}
\name{gtkInputRemove}
\title{gtkInputRemove}
\description{
Removes the function with the given id.
\strong{WARNING: \code{gtk_input_remove} has been deprecated since version 2.4 and should not be used in newly-written code. Use \code{\link{gSourceRemove}} instead.}
}
\usage{gtkInputRemove(input.handler.id)}
\arguments{\item{\verb{input.handler.id}}{Identifies the function to remove.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
