\alias{gtkInvisibleGetScreen}
\name{gtkInvisibleGetScreen}
\title{gtkInvisibleGetScreen}
\description{Returns the \code{\link{GdkScreen}} object associated with \code{invisible}}
\usage{gtkInvisibleGetScreen(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkInvisible}}.}}
\details{Since 2.2}
\value{[\code{\link{GdkScreen}}] the associated \code{\link{GdkScreen}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
