\alias{gtkInvisibleNew}
\name{gtkInvisibleNew}
\title{gtkInvisibleNew}
\description{Creates a new \verb{\link{GtkInvisible}}.}
\usage{gtkInvisibleNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}]  a new \verb{\link{GtkInvisible}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
