\alias{gtkInvisibleNew}
\name{gtkInvisibleNew}
\title{gtkInvisibleNew}
\description{Creates a new \code{\link{GtkInvisible}}.}
\usage{gtkInvisibleNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkInvisible}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
