\alias{gtkInvisibleNewForScreen}
\name{gtkInvisibleNewForScreen}
\title{gtkInvisibleNewForScreen}
\description{Creates a new \code{\link{GtkInvisible}} object for a specified screen}
\usage{gtkInvisibleNewForScreen(screen, show = TRUE)}
\arguments{\item{\verb{screen}}{a \code{\link{GdkScreen}} which identifies on which
the new \code{\link{GtkInvisible}} will be created.}}
\details{Since 2.2}
\value{[\code{\link{GtkWidget}}] a newly created \code{\link{GtkInvisible}} object}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
