\alias{gtkInvisibleSetScreen}
\name{gtkInvisibleSetScreen}
\title{gtkInvisibleSetScreen}
\description{Sets the \code{\link{GdkScreen}} where the \code{\link{GtkInvisible}} object will be displayed.}
\usage{gtkInvisibleSetScreen(object, screen)}
\arguments{
\item{\code{object}}{[\code{\link{GtkInvisible}}]  a \code{\link{GtkInvisible}}.}
\item{\code{screen}}{[\code{\link{GdkScreen}}]  a \code{\link{GdkScreen}}.}
}
\details{  Since  2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
