\alias{gtkInvisibleSetScreen}
\name{gtkInvisibleSetScreen}
\title{gtkInvisibleSetScreen}
\description{Sets the \code{\link{GdkScreen}} where the \code{\link{GtkInvisible}} object will be displayed.}
\usage{gtkInvisibleSetScreen(object, screen)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkInvisible}}.}
\item{\verb{screen}}{a \code{\link{GdkScreen}}.}
}
\details{Since 2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
