\alias{gtkItemDeselect}
\name{gtkItemDeselect}
\title{gtkItemDeselect}
\description{Emits the "deselect" signal on the given item.}
\usage{gtkItemDeselect(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkItem}}] a \verb{\link{GtkItem}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
