\alias{gtkItemFactoryCreateMenuEntries}
\name{gtkItemFactoryCreateMenuEntries}
\title{gtkItemFactoryCreateMenuEntries}
\description{
Creates the menu items from the \code{entries}.
\strong{WARNING: \code{gtk_item_factory_create_menu_entries} has been deprecated since version 2.4 and should not be used in newly-written code. Use \code{\link{GtkUIManager}} instead.}
}
\usage{gtkItemFactoryCreateMenuEntries(entries)}
\arguments{\item{\verb{entries}}{a list of \verb{GtkMenuEntry}s}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
