\alias{gtkItemFactoryDeleteEntries}
\name{gtkItemFactoryDeleteEntries}
\title{gtkItemFactoryDeleteEntries}
\description{
Deletes the menu items which were created from the \code{entries} by the given
item factory.
\strong{WARNING: \code{gtk_item_factory_delete_entries} has been deprecated since version 2.4 and should not be used in newly-written code. }
}
\usage{gtkItemFactoryDeleteEntries(object, entries)}
\arguments{
\item{\code{object}}{[\code{\link{GtkItemFactory}}]  a \code{\link{GtkItemFactory}}}
\item{\code{entries}}{[\code{\link{GtkItemFactoryEntry}}]  a list of \code{\link{GtkItemFactoryEntry}}s }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
