\alias{gtkItemFactoryDeleteEntries}
\name{gtkItemFactoryDeleteEntries}
\title{gtkItemFactoryDeleteEntries}
\description{
Deletes the menu items which were created from the \code{entries} by the given
item factory.
\strong{WARNING: \code{gtk_item_factory_delete_entries} has been deprecated since version 2.4 and should not be used in newly-written code. Use \code{\link{GtkUIManager}} instead.}
}
\usage{gtkItemFactoryDeleteEntries(object, entries)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkItemFactory}}}
\item{\verb{entries}}{a list of \code{\link{GtkItemFactoryEntry}}s}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
