\alias{gtkItemFactoryDeleteEntry}
\name{gtkItemFactoryDeleteEntry}
\title{gtkItemFactoryDeleteEntry}
\description{
Deletes the menu item which was created from \code{entry} by the given
item factory.
\strong{WARNING: \code{gtk_item_factory_delete_entry} is deprecated and should not be used in newly-written code.  
Deletes the menu item which was created from \code{entry} by the given
item factory.  }
}
\usage{gtkItemFactoryDeleteEntry(object, entry)}
\arguments{
\item{\code{object}}{[\code{\link{GtkItemFactory}}]  a \code{\link{GtkItemFactory}}}
\item{\code{entry}}{[\code{\link{GtkItemFactoryEntry}}]  a \code{\link{GtkItemFactoryEntry}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
