\alias{gtkItemFactoryDeleteEntry}
\name{gtkItemFactoryDeleteEntry}
\title{gtkItemFactoryDeleteEntry}
\description{
Deletes the menu item which was created from \code{entry} by the given
item factory.
\strong{WARNING: \code{gtk_item_factory_delete_entry} has been deprecated since version 2.4 and should not be used in newly-written code. Use \code{\link{GtkUIManager}} instead.}
}
\usage{gtkItemFactoryDeleteEntry(object, entry)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkItemFactory}}}
\item{\verb{entry}}{a \code{\link{GtkItemFactoryEntry}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
