\alias{gtkItemFactoryDeleteItem}
\name{gtkItemFactoryDeleteItem}
\title{gtkItemFactoryDeleteItem}
\description{
Deletes the menu item which was created for \code{path} by the given
item factory.
\strong{WARNING: \code{gtk_item_factory_delete_item} has been deprecated since version 2.4 and should not be used in newly-written code. }
}
\usage{gtkItemFactoryDeleteItem(object, path)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkItemFactory}}]  a \verb{\link{GtkItemFactory}}}
\item{\verb{path}}{[character]  a path}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
