\alias{gtkItemFactoryFromWidget}
\name{gtkItemFactoryFromWidget}
\title{gtkItemFactoryFromWidget}
\description{
Obtains the item factory from which a widget was created.
\strong{WARNING: \code{gtk_item_factory_from_widget} is deprecated and should not be used in newly-written code.}
}
\usage{gtkItemFactoryFromWidget(widget)}
\arguments{\item{\code{widget}}{[\code{\link{GtkWidget}}]  a widget}}
\value{[\code{\link{GtkItemFactory}}]  the item factory from which \code{widget} was created, or \code{NULL}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
