\alias{gtkItemSelect}
\name{gtkItemSelect}
\title{gtkItemSelect}
\description{Emits the "select" signal on the given item.}
\usage{gtkItemSelect(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkItem}}] a \verb{\link{GtkItem}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
