\alias{gtkLabelGetAngle}
\name{gtkLabelGetAngle}
\title{gtkLabelGetAngle}
\description{Gets the angle of rotation for the label. See
gtk_label_set_angle.}
\usage{gtkLabelGetAngle(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkLabel}}]  a \code{\link{GtkLabel}}}}
\details{  Since  2.6}
\value{[numeric]  the angle of rotation for the label}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
