\alias{gtkLabelGetLayout}
\name{gtkLabelGetLayout}
\title{gtkLabelGetLayout}
\description{Gets the \verb{\link{PangoLayout}} used to display the label.
The layout is useful to e.g. convert text positions to
pixel positions, in combination with \code{\link{gtkLabelGetLayoutOffsets}}.}
\usage{gtkLabelGetLayout(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkLabel}}]  a \verb{\link{GtkLabel}}}}
\value{[\verb{\link{PangoLayout}}]  the \verb{\link{PangoLayout}} for this label}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
