\alias{gtkLabelGetLineWrapMode}
\name{gtkLabelGetLineWrapMode}
\title{gtkLabelGetLineWrapMode}
\description{Returns line wrap mode used by the label. See \code{\link{gtkLabelSetLineWrapMode}}.}
\usage{gtkLabelGetLineWrapMode(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkLabel}}}}
\details{Since 2.10}
\value{[\code{\link{PangoWrapMode}}] \code{TRUE} if the lines of the label are automatically wrapped.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
