\alias{gtkLabelGetSelectable}
\name{gtkLabelGetSelectable}
\title{gtkLabelGetSelectable}
\description{Gets the value set by \code{\link{gtkLabelSetSelectable}}.}
\usage{gtkLabelGetSelectable(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkLabel}}}}
\value{[logical] \code{TRUE} if the user can copy text from the label}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
