\alias{gtkLabelSetEllipsize}
\name{gtkLabelSetEllipsize}
\title{gtkLabelSetEllipsize}
\description{Sets the mode used to ellipsize (add an ellipsis: "...") to the text 
if there is not enough space to render the entire string.}
\usage{gtkLabelSetEllipsize(object, mode)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkLabel}}}
\item{\verb{mode}}{a \code{\link{PangoEllipsizeMode}}}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
