\alias{gtkLabelSetLabel}
\name{gtkLabelSetLabel}
\title{gtkLabelSetLabel}
\description{Sets the text of the label. The label is interpreted as
including embedded underlines and/or Pango markup depending
on the values of the \verb{"use-underline"}" and
\verb{"use-markup"} properties.}
\usage{gtkLabelSetLabel(object, str)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkLabel}}}
\item{\verb{str}}{the new text to set for the label}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
