\alias{gtkLabelSetSelectable}
\name{gtkLabelSetSelectable}
\title{gtkLabelSetSelectable}
\description{Selectable labels allow the user to select text from the label, for
copy-and-paste.}
\usage{gtkLabelSetSelectable(object, setting)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkLabel}}}
\item{\verb{setting}}{\code{TRUE} to allow selecting text in the label}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
