\alias{gtkLabelSetText}
\name{gtkLabelSetText}
\title{gtkLabelSetText}
\description{Sets the text within the \code{\link{GtkLabel}} widget. It overwrites any text that
was there before.  }
\usage{gtkLabelSetText(object, str)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkLabel}}}
\item{\verb{str}}{The text you want to set}
}
\details{This will also clear any previously set mnemonic accelerators.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
