\alias{gtkLabelSetUseMarkup}
\name{gtkLabelSetUseMarkup}
\title{gtkLabelSetUseMarkup}
\description{Sets whether the text of the label contains markup in Pango's text markup
language. See \code{\link{gtkLabelSetMarkup}}.}
\usage{gtkLabelSetUseMarkup(object, setting)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkLabel}}}
\item{\verb{setting}}{\code{TRUE} if the label's text should be parsed for markup.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
