\alias{gtkLabelSetUseUnderline}
\name{gtkLabelSetUseUnderline}
\title{gtkLabelSetUseUnderline}
\description{If true, an underline in the text indicates the next character should be
used for the mnemonic accelerator key.}
\usage{gtkLabelSetUseUnderline(object, setting)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkLabel}}}
\item{\verb{setting}}{\code{TRUE} if underlines in the text indicate mnemonics}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
