\alias{gtkLabelSetWidthChars}
\name{gtkLabelSetWidthChars}
\title{gtkLabelSetWidthChars}
\description{Sets the desired width in characters of \code{label} to \code{n.chars}.}
\usage{gtkLabelSetWidthChars(object, n.chars)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkLabel}}}
\item{\verb{n.chars}}{the new desired width, in characters.}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
