\alias{gtkLayoutMove}
\name{gtkLayoutMove}
\title{gtkLayoutMove}
\description{Moves a current child of \code{layout} to a new position.}
\usage{gtkLayoutMove(object, child.widget, x, y)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkLayout}}]  a \verb{\link{GtkLayout}}}
\item{\verb{child.widget}}{[\verb{\link{GtkWidget}}]  a current child of \code{layout}}
\item{\verb{x}}{[integer]  X position to move to}
\item{\verb{y}}{[integer]  Y position to move to}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
