\alias{gtkLayoutMove}
\name{gtkLayoutMove}
\title{gtkLayoutMove}
\description{Moves a current child of \code{layout} to a new position.}
\usage{gtkLayoutMove(object, child.widget, x, y)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkLayout}}}
\item{\verb{child.widget}}{a current child of \code{layout}}
\item{\verb{x}}{X position to move to}
\item{\verb{y}}{Y position to move to}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
