\alias{gtkLayoutPut}
\name{gtkLayoutPut}
\title{gtkLayoutPut}
\description{Adds \code{child.widget} to \code{layout}, at position (\code{x},\code{y}).
\code{layout} becomes the new parent container of \code{child.widget}.}
\usage{gtkLayoutPut(object, child.widget, x, y)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkLayout}}}
\item{\verb{child.widget}}{child widget}
\item{\verb{x}}{X position of child widget}
\item{\verb{y}}{Y position of child widget}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
