\alias{gtkLayoutSetHadjustment}
\name{gtkLayoutSetHadjustment}
\title{gtkLayoutSetHadjustment}
\description{Sets the horizontal scroll adjustment for the layout.}
\usage{gtkLayoutSetHadjustment(object, adjustment = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkLayout}}]  a \verb{\link{GtkLayout}}}
\item{\verb{adjustment}}{[\verb{\link{GtkAdjustment}}]  new scroll adjustment}
}
\details{See \verb{\link{GtkScrolledWindow}}, \verb{\link{GtkScrollbar}}, \verb{\link{GtkAdjustment}} for details.  }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
