\alias{gtkLayoutSetSize}
\name{gtkLayoutSetSize}
\title{gtkLayoutSetSize}
\description{Sets the size of the scrollable area of the layout.}
\usage{gtkLayoutSetSize(object, width, height)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkLayout}}}
\item{\verb{width}}{width of entire scrollable area}
\item{\verb{height}}{height of entire scrollable area}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
