\alias{gtkLayoutSetVadjustment}
\name{gtkLayoutSetVadjustment}
\title{gtkLayoutSetVadjustment}
\description{Sets the vertical scroll adjustment for the layout.}
\usage{gtkLayoutSetVadjustment(object, adjustment = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkLayout}}}
\item{\verb{adjustment}}{new scroll adjustment. \emph{[  \acronym{allow-none}  ]}}
}
\details{See \code{\link{GtkScrolledWindow}}, \code{\link{GtkScrollbar}}, \code{\link{GtkAdjustment}} for details.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
