\alias{gtkLinkButtonNew}
\name{gtkLinkButtonNew}
\title{gtkLinkButtonNew}
\description{Creates a new \code{\link{GtkLinkButton}} with the URI as its text.}
\usage{gtkLinkButtonNew(uri)}
\arguments{\item{\verb{uri}}{a valid URI}}
\details{Since 2.10}
\value{[\code{\link{GtkWidget}}] a new link button widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
