\alias{gtkLinkButtonNewWithLabel}
\name{gtkLinkButtonNewWithLabel}
\title{gtkLinkButtonNewWithLabel}
\description{Creates a new \code{\link{GtkLinkButton}} containing a label.}
\usage{gtkLinkButtonNewWithLabel(uri, label = NULL, show = TRUE)}
\arguments{
\item{\verb{uri}}{a valid URI}
\item{\verb{label}}{the text of the button. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.10}
\value{[\code{\link{GtkWidget}}] a new link button widget. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
