\alias{gtkLinkButtonSetUri}
\name{gtkLinkButtonSetUri}
\title{gtkLinkButtonSetUri}
\description{Sets \code{uri} as the URI where the \code{\link{GtkLinkButton}} points. As a side-effect
this unsets the 'visited' state of the button.}
\usage{gtkLinkButtonSetUri(object, uri)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkLinkButton}}}
\item{\verb{uri}}{a valid URI}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
