\alias{gtkListAppendItems}
\name{gtkListAppendItems}
\title{gtkListAppendItems}
\description{
Adds \code{items} to the end of the \code{list}.
\strong{WARNING: \code{gtk_list_append_items} is deprecated and should not be used in newly-written code.  
Adds \code{items} to the end of the \code{list}.}
}
\usage{gtkListAppendItems(object, items)}
\arguments{
\item{\code{object}}{[\code{\link{GtkList}}] the list widget.}
\item{\code{items}}{[list] the items.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
