\alias{gtkListEndDragSelection}
\name{gtkListEndDragSelection}
\title{gtkListEndDragSelection}
\description{
Stops the drag selection mode and ungrabs the pointer. This has no
effect if a drag selection is not active.
\strong{WARNING: \code{gtk_list_end_drag_selection} is deprecated and should not be used in newly-written code.}
}
\usage{gtkListEndDragSelection(object)}
\arguments{\item{\verb{object}}{the list widget.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
