\alias{gtkListInsertItems}
\name{gtkListInsertItems}
\title{gtkListInsertItems}
\description{
Inserts \code{items} into the \code{list} at the position \code{position}.
\strong{WARNING: \code{gtk_list_insert_items} is deprecated and should not be used in newly-written code.}
}
\usage{gtkListInsertItems(object, items, position)}
\arguments{
\item{\verb{object}}{the list widget.}
\item{\verb{items}}{the items.}
\item{\verb{position}}{the position to insert \code{items}, starting at 0.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
