\alias{gtkListItemNew}
\name{gtkListItemNew}
\title{gtkListItemNew}
\description{
Creates a new \verb{GtkListitem}.
\strong{WARNING: \code{gtk_list_item_new} is deprecated and should not be used in newly-written code.}
}
\usage{gtkListItemNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkListItem}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
