\alias{gtkListItemNewWithLabel}
\name{gtkListItemNewWithLabel}
\title{gtkListItemNewWithLabel}
\description{
Creates a new \code{\link{GtkListItem}} with a child label containing the given string.
\strong{WARNING: \code{gtk_list_item_new_with_label} is deprecated and should not be used in newly-written code.}
}
\usage{gtkListItemNewWithLabel(label, show = TRUE)}
\arguments{\item{\verb{label}}{the string to use for the child label.}}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkListItem}} with a child \code{\link{GtkLabel}} with the text set to
\code{label}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
