\alias{gtkListItemSelect}
\name{gtkListItemSelect}
\title{gtkListItemSelect}
\description{
Selects the item, by emitting the item's "select" signal.
Depending on the selection mode of the list, this may cause other items to
be deselected.
\strong{WARNING: \code{gtk_list_item_select} is deprecated and should not be used in newly-written code.}
}
\usage{gtkListItemSelect(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkListItem}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
