\alias{gtkListRemoveItems}
\name{gtkListRemoveItems}
\title{gtkListRemoveItems}
\description{
Removes the \code{items} from the \code{list}.
\strong{WARNING: \code{gtk_list_remove_items} is deprecated and should not be used in newly-written code.}
}
\usage{gtkListRemoveItems(object, items)}
\arguments{
\item{\verb{object}}{the list widget.}
\item{\verb{items}}{the items to remove.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
