\alias{gtkListSelectChild}
\name{gtkListSelectChild}
\title{gtkListSelectChild}
\description{
Selects the given \code{child}. The signal GtkList::select-child will be
emitted.
\strong{WARNING: \code{gtk_list_select_child} is deprecated and should not be used in newly-written code.}
}
\usage{gtkListSelectChild(object, child)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkList}}] the list widget}
\item{\verb{child}}{[\verb{\link{GtkWidget}}] the child to select.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
