\alias{gtkListStartSelection}
\name{gtkListStartSelection}
\title{gtkListStartSelection}
\description{
Starts a selection (or part of selection) at the focused child. Only in
\verb{GTK_SELECTION_EXTENDED} mode.
\strong{WARNING: \code{gtk_list_start_selection} is deprecated and should not be used in newly-written code.}
}
\usage{gtkListStartSelection(object)}
\arguments{\item{\verb{object}}{the list widget.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
