\alias{gtkListStoreClear}
\name{gtkListStoreClear}
\title{gtkListStoreClear}
\description{Removes all rows from the list store.}
\usage{gtkListStoreClear(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkListStore}}]  a \verb{\link{GtkListStore}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
