\alias{gtkListStoreIterIsValid}
\name{gtkListStoreIterIsValid}
\title{gtkListStoreIterIsValid}
\description{\strong{WARNING: }}
\usage{gtkListStoreIterIsValid(object, iter)}
\arguments{
\item{\code{object}}{[\code{\link{GtkListStore}}]  A \code{\link{GtkListStore}}.}
\item{\code{iter}}{[\code{\link{GtkTreeIter}}]  A \code{\link{GtkTreeIter}}.}
}
\details{Checks if the given iter is a valid iter for this \code{\link{GtkListStore}}.  

  Since  2.2}
\value{[logical]  \code{TRUE} if the iter is valid, \code{FALSE} if the iter is invalid.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
