\alias{gtkListStoreNewv}
\name{gtkListStoreNewv}
\title{gtkListStoreNewv}
\description{Non-vararg creation function.  Used primarily by language bindings.}
\usage{gtkListStoreNewv(value)}
\arguments{\item{\code{value}}{[\code{\link{GType}}]  a list of \code{\link{GType}} types for the columns, from first to last}}
\value{[\code{\link{GtkListStore}}]  a new \code{\link{GtkListStore}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
