\alias{gtkListStoreNewv}
\name{gtkListStoreNewv}
\title{gtkListStoreNewv}
\description{Non-vararg creation function.  Used primarily by language bindings.}
\usage{gtkListStoreNewv(value)}
\arguments{\item{\verb{value}}{a list of \code{\link{GType}} types for the columns, from first to last. \emph{[  \acronym{array}   length=n_columns]}}}
\value{[\code{\link{GtkListStore}}] a new \code{\link{GtkListStore}}. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
