\alias{gtkListStoreSwap}
\name{gtkListStoreSwap}
\title{gtkListStoreSwap}
\description{Swaps \code{a} and \code{b} in \code{store}. Note that this function only works with
unsorted stores.}
\usage{gtkListStoreSwap(object, a, b)}
\arguments{
\item{\code{object}}{[\code{\link{GtkListStore}}]  A \code{\link{GtkListStore}}.}
\item{\code{a}}{[\code{\link{GtkTreeIter}}]  A \code{\link{GtkTreeIter}}.}
\item{\code{b}}{[\code{\link{GtkTreeIter}}]  Another \code{\link{GtkTreeIter}}.}
}
\details{  Since  2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
