\alias{gtkListStoreSwap}
\name{gtkListStoreSwap}
\title{gtkListStoreSwap}
\description{Swaps \code{a} and \code{b} in \code{store}. Note that this function only works with
unsorted stores.}
\usage{gtkListStoreSwap(object, a, b)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkListStore}}.}
\item{\verb{a}}{A \code{\link{GtkTreeIter}}.}
\item{\verb{b}}{Another \code{\link{GtkTreeIter}}.}
}
\details{Since 2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
