\alias{gtkListToggleFocusRow}
\name{gtkListToggleFocusRow}
\title{gtkListToggleFocusRow}
\description{
Toggles the focus row. If the focus row is selected, it's
unselected. If the focus row is unselected, it's selected. If the
selection mode of \code{list} is \verb{GTK_SELECTION_BROWSE}, this has no effect,
as the selection is always at the focus row.
\strong{WARNING: \code{gtk_list_toggle_focus_row} is deprecated and should not be used in newly-written code.}
}
\usage{gtkListToggleFocusRow(object)}
\arguments{\item{\verb{object}}{the list widget.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
