\alias{gtkListUnselectChild}
\name{gtkListUnselectChild}
\title{gtkListUnselectChild}
\description{
Unselects the given \code{child}. The signal GtkList::unselect-child will be
emitted.
\strong{WARNING: \code{gtk_list_unselect_child} is deprecated and should not be used in newly-written code.}
}
\usage{gtkListUnselectChild(object, child)}
\arguments{
\item{\verb{object}}{the list widget.}
\item{\verb{child}}{the child to unselect.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
