\alias{gtkMain}
\name{gtkMain}
\title{gtkMain}
\description{Runs the main loop until \code{\link{gtkMainQuit}} is called. You can nest calls to
\code{\link{gtkMain}}. In that case \code{\link{gtkMainQuit}} will make the innermost invocation
of the main loop return.}
\usage{gtkMain()}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
