\alias{gtkMainLevel}
\name{gtkMainLevel}
\title{gtkMainLevel}
\description{Asks for the current nesting level of the main loop. This can be useful
when calling \code{\link{gtkQuitAdd}}.}
\usage{gtkMainLevel()}
\value{[numeric] the nesting level of the current invocation of the main loop.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
