\alias{gtkMenuBarSetChildPackDirection}
\name{gtkMenuBarSetChildPackDirection}
\title{gtkMenuBarSetChildPackDirection}
\description{Sets how widgets should be packed inside the children of a menubar.}
\usage{gtkMenuBarSetChildPackDirection(object, child.pack.dir)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkMenuBar}}}
\item{\verb{child.pack.dir}}{a new \code{\link{GtkPackDirection}}}
}
\details{Since 2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
