\alias{gtkMenuBarSetPackDirection}
\name{gtkMenuBarSetPackDirection}
\title{gtkMenuBarSetPackDirection}
\description{Sets how items should be packed inside a menubar.}
\usage{gtkMenuBarSetPackDirection(object, pack.dir)}
\arguments{
\item{\code{object}}{[\code{\link{GtkMenuBar}}]  a \code{\link{GtkMenuBar}}.}
\item{\code{pack.dir}}{[\code{\link{GtkPackDirection}}]  a new \code{\link{GtkPackDirection}}.}
}
\details{  Since  2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
