\alias{gtkMenuGetAccelGroup}
\name{gtkMenuGetAccelGroup}
\title{gtkMenuGetAccelGroup}
\description{Gets the \code{\link{GtkAccelGroup}} which holds global accelerators for the menu.
See \code{\link{gtkMenuSetAccelGroup}}.}
\usage{gtkMenuGetAccelGroup(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMenu}}.}}
\value{[\code{\link{GtkAccelGroup}}] the \code{\link{GtkAccelGroup}} associated with the menu.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
