\alias{gtkMenuGetAttachWidget}
\name{gtkMenuGetAttachWidget}
\title{gtkMenuGetAttachWidget}
\description{Returns the \code{\link{GtkWidget}} that the menu is attached to.}
\usage{gtkMenuGetAttachWidget(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMenu}}.}}
\value{[\code{\link{GtkWidget}}] the \code{\link{GtkWidget}} that the menu is attached to.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
