\alias{gtkMenuGetForAttachWidget}
\name{gtkMenuGetForAttachWidget}
\title{gtkMenuGetForAttachWidget}
\usage{gtkMenuGetForAttachWidget(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkWidget}}]  a \code{\link{GtkWidget}}
* Returns a list of the menus which are attached to this widget.
This list is owned by GTK+ and must not be modified.}}
\details{  Since  2.6}
\value{[list]  the list of menus attached to his widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
