\alias{gtkMenuGetForAttachWidget}
\name{gtkMenuGetForAttachWidget}
\title{gtkMenuGetForAttachWidget}
\description{Returns a list of the menus which are attached to this widget.}
\usage{gtkMenuGetForAttachWidget(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.6}
\value{[list] the list of menus attached to his widget. \emph{[  \acronym{element-type}   GtkWidget][  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
