\alias{gtkMenuGetTearoffState}
\name{gtkMenuGetTearoffState}
\title{gtkMenuGetTearoffState}
\description{Returns whether the menu is torn off. See
\code{\link{gtkMenuSetTearoffState}}.}
\usage{gtkMenuGetTearoffState(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMenu}}}}
\value{[logical] \code{TRUE} if the menu is currently torn off.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
