\alias{gtkMenuItemActivate}
\name{gtkMenuItemActivate}
\title{gtkMenuItemActivate}
\description{Emits the "activate" signal on the given item}
\usage{gtkMenuItemActivate(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkMenuItem}}] the menu item}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
