\alias{gtkMenuItemDeselect}
\name{gtkMenuItemDeselect}
\title{gtkMenuItemDeselect}
\description{Emits the "deselect" signal on the given item. Behaves exactly like
\code{\link{gtkItemDeselect}}.}
\usage{gtkMenuItemDeselect(object)}
\arguments{\item{\verb{object}}{the menu item}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
