\alias{gtkMenuItemGetSubmenu}
\name{gtkMenuItemGetSubmenu}
\title{gtkMenuItemGetSubmenu}
\description{Gets the submenu underneath this menu item, if any. See
\code{\link{gtkMenuItemSetSubmenu}}.}
\usage{gtkMenuItemGetSubmenu(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMenuItem}}}}
\value{[\code{\link{GtkWidget}}] submenu for this menu item, or \code{NULL} if none.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
