\alias{gtkMenuItemNew}
\name{gtkMenuItemNew}
\title{gtkMenuItemNew}
\description{Creates a new \code{\link{GtkMenuItem}}.}
\usage{gtkMenuItemNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] the newly created \code{\link{GtkMenuItem}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
