\alias{gtkMenuItemNewWithLabel}
\name{gtkMenuItemNewWithLabel}
\title{gtkMenuItemNewWithLabel}
\description{Creates a new \code{\link{GtkMenuItem}} whose child is a \code{\link{GtkLabel}}.}
\usage{gtkMenuItemNewWithLabel(label, show = TRUE)}
\arguments{\item{\verb{label}}{the text for the label}}
\value{[\code{\link{GtkWidget}}] the newly created \code{\link{GtkMenuItem}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
