\alias{gtkMenuItemRemoveSubmenu}
\name{gtkMenuItemRemoveSubmenu}
\title{gtkMenuItemRemoveSubmenu}
\description{
Removes the widget's submenu.
\strong{WARNING: \code{gtk_menu_item_remove_submenu} has been deprecated since version 2.12 and should not be used in newly-written code. \code{\link{gtkMenuItemRemoveSubmenu}} is deprecated and
                  should not be used in newly written code. Use
                  \code{\link{gtkMenuItemSetSubmenu}} instead.}
}
\usage{gtkMenuItemRemoveSubmenu(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMenuItem}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
