\alias{gtkMenuItemSelect}
\name{gtkMenuItemSelect}
\title{gtkMenuItemSelect}
\description{Emits the "select" signal on the given item. Behaves exactly like
\code{\link{gtkItemSelect}}.}
\usage{gtkMenuItemSelect(object)}
\arguments{\item{\verb{object}}{the menu item}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
