\alias{gtkMenuItemSetSubmenu}
\name{gtkMenuItemSetSubmenu}
\title{gtkMenuItemSetSubmenu}
\description{Sets or replaces the menu item's submenu, or removes it when a \code{NULL}
submenu is passed.}
\usage{gtkMenuItemSetSubmenu(object, submenu)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkMenuItem}}]  a \verb{\link{GtkMenuItem}}}
\item{\verb{submenu}}{[\verb{\link{GtkWidget}}]  the submenu, or \code{NULL}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
