\alias{gtkMenuItemSetSubmenu}
\name{gtkMenuItemSetSubmenu}
\title{gtkMenuItemSetSubmenu}
\description{Sets or replaces the menu item's submenu, or removes it when a \code{NULL}
submenu is passed.}
\usage{gtkMenuItemSetSubmenu(object, submenu)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkMenuItem}}}
\item{\verb{submenu}}{the submenu, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
