\alias{gtkMenuItemToggleSizeAllocate}
\name{gtkMenuItemToggleSizeAllocate}
\title{gtkMenuItemToggleSizeAllocate}
\description{Emits the "toggle_size_allocate" signal on the given item.}
\usage{gtkMenuItemToggleSizeAllocate(object, allocation)}
\arguments{
\item{\verb{object}}{the menu item.}
\item{\verb{allocation}}{the allocation to use as signal data.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
