\alias{gtkMenuItemToggleSizeRequest}
\name{gtkMenuItemToggleSizeRequest}
\title{gtkMenuItemToggleSizeRequest}
\description{Emits the "toggle_size_request" signal on the given item.}
\usage{gtkMenuItemToggleSizeRequest(object, requisition)}
\arguments{
\item{\verb{object}}{the menu item}
\item{\verb{requisition}}{the requisition to use as signal data.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
