\alias{gtkMenuPopdown}
\name{gtkMenuPopdown}
\title{gtkMenuPopdown}
\description{Removes the menu from the screen.}
\usage{gtkMenuPopdown(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkMenu}}] a \verb{\link{GtkMenu}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
