\alias{gtkMenuReorderChild}
\name{gtkMenuReorderChild}
\title{gtkMenuReorderChild}
\description{Moves a \code{\link{GtkMenuItem}} to a new position within the \code{\link{GtkMenu}}.}
\usage{gtkMenuReorderChild(object, child, position)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkMenu}}.}
\item{\verb{child}}{the \code{\link{GtkMenuItem}} to move.}
\item{\verb{position}}{the new position to place \code{child}.  Positions are numbered from
0 to n-1.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
