\alias{gtkMenuSetScreen}
\name{gtkMenuSetScreen}
\title{gtkMenuSetScreen}
\description{Sets the \code{\link{GdkScreen}} on which the menu will be displayed.}
\usage{gtkMenuSetScreen(object, screen = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkMenu}}.}
\item{\verb{screen}}{a \code{\link{GdkScreen}}, or \code{NULL} if the screen should be
determined by the widget the menu is attached to. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
