\alias{gtkMenuShellActivateItem}
\name{gtkMenuShellActivateItem}
\title{gtkMenuShellActivateItem}
\description{Activates the menu item within the menu shell.}
\usage{gtkMenuShellActivateItem(object, menu.item, force.deactivate)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkMenuShell}}] a \verb{\link{GtkMenuShell}}.}
\item{\verb{menu.item}}{[\verb{\link{GtkWidget}}] The \verb{\link{GtkMenuItem}} to activate.}
\item{\verb{force.deactivate}}{[logical] If TRUE, force the deactivation of the menu shell
after the menu item is activated.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
