\alias{gtkMenuShellAppend}
\name{gtkMenuShellAppend}
\title{gtkMenuShellAppend}
\description{Adds a new \verb{\link{GtkMenuItem}} to the end of the menu shell's item list.}
\usage{gtkMenuShellAppend(object, child)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkMenuShell}}] a \verb{\link{GtkMenuShell}}.}
\item{\verb{child}}{[\verb{\link{GtkWidget}}] The \verb{\link{GtkMenuItem}} to add.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
