\alias{gtkMenuShellAppend}
\name{gtkMenuShellAppend}
\title{gtkMenuShellAppend}
\description{Adds a new \code{\link{GtkMenuItem}} to the end of the menu shell's item list.}
\usage{gtkMenuShellAppend(object, child)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkMenuShell}}.}
\item{\verb{child}}{The \code{\link{GtkMenuItem}} to add.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
