\alias{gtkMenuShellDeactivate}
\name{gtkMenuShellDeactivate}
\title{gtkMenuShellDeactivate}
\description{Deactivates the menu shell.  Typically this results in the menu shell
being erased from the screen.}
\usage{gtkMenuShellDeactivate(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMenuShell}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
