\alias{gtkMenuShellGetTakeFocus}
\name{gtkMenuShellGetTakeFocus}
\title{gtkMenuShellGetTakeFocus}
\description{Returns \code{TRUE} if the menu shell will take the keyboard focus on popup.}
\usage{gtkMenuShellGetTakeFocus(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMenuShell}}}}
\details{Since 2.8}
\value{[logical] \code{TRUE} if the menu shell will take the keyboard focus on popup.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
