\alias{gtkMenuShellInsert}
\name{gtkMenuShellInsert}
\title{gtkMenuShellInsert}
\description{Adds a new \code{\link{GtkMenuItem}} to the menu shell's item list at the position
indicated by \code{position}. }
\usage{gtkMenuShellInsert(object, child, position)}
\arguments{
\item{\code{object}}{[\code{\link{GtkMenuShell}}] a \code{\link{GtkMenuShell}}.}
\item{\code{child}}{[\code{\link{GtkWidget}}] The \code{\link{GtkMenuItem}} to add.}
\item{\code{position}}{[integer] The position in the item list where \code{child} is added.
Positions are numbered from 0 to n-1.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
