\alias{gtkMenuShellInsert}
\name{gtkMenuShellInsert}
\title{gtkMenuShellInsert}
\description{Adds a new \code{\link{GtkMenuItem}} to the menu shell's item list at the position
indicated by \code{position}.}
\usage{gtkMenuShellInsert(object, child, position)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkMenuShell}}.}
\item{\verb{child}}{The \code{\link{GtkMenuItem}} to add.}
\item{\verb{position}}{The position in the item list where \code{child} is added.
Positions are numbered from 0 to n-1.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
