\alias{gtkMenuShellPrepend}
\name{gtkMenuShellPrepend}
\title{gtkMenuShellPrepend}
\description{Adds a new \code{\link{GtkMenuItem}} to the beginning of the menu shell's item list.}
\usage{gtkMenuShellPrepend(object, child)}
\arguments{
\item{\code{object}}{[\code{\link{GtkMenuShell}}] a \code{\link{GtkMenuShell}}.}
\item{\code{child}}{[\code{\link{GtkWidget}}] The \code{\link{GtkMenuItem}} to add.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
