\alias{gtkMenuShellSelectItem}
\name{gtkMenuShellSelectItem}
\title{gtkMenuShellSelectItem}
\description{Selects the menu item from the menu shell.}
\usage{gtkMenuShellSelectItem(object, menu.item)}
\arguments{
\item{\code{object}}{[\code{\link{GtkMenuShell}}] a \code{\link{GtkMenuShell}}.}
\item{\code{menu.item}}{[\code{\link{GtkWidget}}] The \code{\link{GtkMenuItem}} to select.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
