\alias{gtkMenuShellSelectItem}
\name{gtkMenuShellSelectItem}
\title{gtkMenuShellSelectItem}
\description{Selects the menu item from the menu shell.}
\usage{gtkMenuShellSelectItem(object, menu.item)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkMenuShell}}.}
\item{\verb{menu.item}}{The \code{\link{GtkMenuItem}} to select.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
