\alias{gtkMenuToolButtonGetMenu}
\name{gtkMenuToolButtonGetMenu}
\title{gtkMenuToolButtonGetMenu}
\description{Gets the \code{\link{GtkMenu}} associated with \code{\link{GtkMenuToolButton}}.}
\usage{gtkMenuToolButtonGetMenu(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMenuToolButton}}}}
\details{Since 2.6}
\value{[\code{\link{GtkWidget}}] the \code{\link{GtkMenu}} associated with \code{\link{GtkMenuToolButton}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
