\alias{gtkMenuToolButtonNew}
\name{gtkMenuToolButtonNew}
\title{gtkMenuToolButtonNew}
\description{Creates a new \verb{\link{GtkMenuToolButton}} using \code{icon.widget} as icon and
\code{label} as label.}
\usage{gtkMenuToolButtonNew(icon.widget, label, show = TRUE)}
\arguments{
\item{\verb{icon.widget}}{[\verb{\link{GtkWidget}}]  a widget that will be used as icon widget, or \code{NULL}}
\item{\verb{label}}{[character]  a string that will be used as label, or \code{NULL}}
}
\details{  Since  2.6}
\value{[\verb{\link{GtkToolItem}}]  the new \verb{\link{GtkMenuToolButton}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
