\alias{gtkMenuToolButtonNew}
\name{gtkMenuToolButtonNew}
\title{gtkMenuToolButtonNew}
\description{Creates a new \code{\link{GtkMenuToolButton}} using \code{icon.widget} as icon and
\code{label} as label.}
\usage{gtkMenuToolButtonNew(icon.widget, label, show = TRUE)}
\arguments{
\item{\verb{icon.widget}}{a widget that will be used as icon widget, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
\item{\verb{label}}{a string that will be used as label, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.6}
\value{[\code{\link{GtkToolItem}}] the new \code{\link{GtkMenuToolButton}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
