\alias{gtkMenuToolButtonNewFromStock}
\name{gtkMenuToolButtonNewFromStock}
\title{gtkMenuToolButtonNewFromStock}
\description{Creates a new \code{\link{GtkMenuToolButton}}.
The new \code{\link{GtkMenuToolButton}} will contain an icon and label from
the stock item indicated by \code{stock.id}.}
\usage{gtkMenuToolButtonNewFromStock(stock.id)}
\arguments{\item{\verb{stock.id}}{the name of a stock item}}
\details{Since 2.6}
\value{[\code{\link{GtkToolItem}}] the new \code{\link{GtkMenuToolButton}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
