\alias{gtkMenuToolButtonSetArrowTooltipText}
\name{gtkMenuToolButtonSetArrowTooltipText}
\title{gtkMenuToolButtonSetArrowTooltipText}
\description{Sets the tooltip text to be used as tooltip for the arrow button which
pops up the menu.  See \code{\link{gtkToolItemSetTooltip}} for setting a tooltip
on the whole \code{\link{GtkMenuToolButton}}.}
\usage{gtkMenuToolButtonSetArrowTooltipText(object, text)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkMenuToolButton}}}
\item{\verb{text}}{text to be used as tooltip text for button's arrow button}
}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
