\alias{gtkMenuToolButtonSetMenu}
\name{gtkMenuToolButtonSetMenu}
\title{gtkMenuToolButtonSetMenu}
\description{Sets the \code{\link{GtkMenu}} that is popped up when the user clicks on the arrow.
If \code{menu} is NULL, the arrow button becomes insensitive.}
\usage{gtkMenuToolButtonSetMenu(object, menu)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkMenuToolButton}}}
\item{\verb{menu}}{the \code{\link{GtkMenu}} associated with \code{\link{GtkMenuToolButton}}}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
