\alias{gtkMessageDialogSetImage}
\name{gtkMessageDialogSetImage}
\title{gtkMessageDialogSetImage}
\description{Sets the dialog's image to \code{image}.}
\usage{gtkMessageDialogSetImage(object, image)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkMessageDialog}}}
\item{\verb{image}}{the image}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
