\alias{gtkMiscGetAlignment}
\name{gtkMiscGetAlignment}
\title{gtkMiscGetAlignment}
\description{Gets the X and Y alignment of the widget within its allocation. 
See \code{\link{gtkMiscSetAlignment}}.}
\usage{gtkMiscGetAlignment(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkMisc}}]  a \verb{\link{GtkMisc}}}}
\value{
A list containing the following elements:
\item{\verb{xalign}}{[numeric]  location to store X alignment of \code{misc}, or \code{NULL}}
\item{\verb{yalign}}{[numeric]  location to store Y alignment of \code{misc}, or \code{NULL}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
