\alias{gtkMiscGetPadding}
\name{gtkMiscGetPadding}
\title{gtkMiscGetPadding}
\description{Gets the padding in the X and Y directions of the widget. 
See \code{\link{gtkMiscSetPadding}}.}
\usage{gtkMiscGetPadding(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkMisc}}}}
\value{
A list containing the following elements:
\item{\verb{xpad}}{location to store padding in the X direction, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
\item{\verb{ypad}}{location to store padding in the Y direction, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
