\alias{gtkMiscSetPadding}
\name{gtkMiscSetPadding}
\title{gtkMiscSetPadding}
\description{Sets the amount of space to add around the widget.}
\usage{gtkMiscSetPadding(object, xpad, ypad)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkMisc}}.}
\item{\verb{xpad}}{the amount of space to add on the left and right of the widget,
in pixels.}
\item{\verb{ypad}}{the amount of space to add on the top and bottom of the widget,
in pixels.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
