\alias{gtkNotebookGetCurrentPage}
\name{gtkNotebookGetCurrentPage}
\title{gtkNotebookGetCurrentPage}
\description{Returns the page number of the current page.}
\usage{gtkNotebookGetCurrentPage(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkNotebook}}}}
\value{[integer] the index (starting from 0) of the current
page in the notebook. If the notebook has no pages, then
-1 will be returned.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
