\alias{gtkNotebookGetNPages}
\name{gtkNotebookGetNPages}
\title{gtkNotebookGetNPages}
\description{Gets the number of pages in a notebook.}
\usage{gtkNotebookGetNPages(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkNotebook}}]  a \verb{\link{GtkNotebook}}}}
\details{  Since  2.2}
\value{[integer]  the number of pages in the notebook.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
