\alias{gtkNotebookGetScrollable}
\name{gtkNotebookGetScrollable}
\title{gtkNotebookGetScrollable}
\description{Returns whether the tab label area has arrows for scrolling. See
\code{\link{gtkNotebookSetScrollable}}.}
\usage{gtkNotebookGetScrollable(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkNotebook}}]  a \code{\link{GtkNotebook}}}}
\value{[logical] \code{TRUE} if arrows for scrolling are present}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
