\alias{gtkNotebookGetShowBorder}
\name{gtkNotebookGetShowBorder}
\title{gtkNotebookGetShowBorder}
\description{Returns whether a bevel will be drawn around the notebook pages. See
\code{\link{gtkNotebookSetShowBorder}}.}
\usage{gtkNotebookGetShowBorder(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkNotebook}}}}
\value{[logical] \code{TRUE} if the bevel is drawn}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
