\alias{gtkNotebookGetShowTabs}
\name{gtkNotebookGetShowTabs}
\title{gtkNotebookGetShowTabs}
\description{Returns whether the tabs of the notebook are shown. See
\code{\link{gtkNotebookSetShowTabs}}.}
\usage{gtkNotebookGetShowTabs(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkNotebook}}}}
\value{[logical] \code{TRUE} if the tabs are shown}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
