\alias{gtkNotebookGetTabDetachable}
\name{gtkNotebookGetTabDetachable}
\title{gtkNotebookGetTabDetachable}
\description{Returns whether the tab contents can be detached from \code{notebook}.}
\usage{gtkNotebookGetTabDetachable(object, child)}
\arguments{
\item{\code{object}}{[\code{\link{GtkNotebook}}]  a \code{\link{GtkNotebook}}}
\item{\code{child}}{[\code{\link{GtkWidget}}]  a child \code{\link{GtkWidget}}}
}
\details{  Since  2.10}
\value{[logical]  TRUE if the tab is detachable.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
