\alias{gtkNotebookGetTabDetachable}
\name{gtkNotebookGetTabDetachable}
\title{gtkNotebookGetTabDetachable}
\description{Returns whether the tab contents can be detached from \code{notebook}.}
\usage{gtkNotebookGetTabDetachable(object, child)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkNotebook}}}
\item{\verb{child}}{a child \code{\link{GtkWidget}}}
}
\details{Since 2.10}
\value{[logical] TRUE if the tab is detachable.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
