\alias{gtkNotebookGetTabReorderable}
\name{gtkNotebookGetTabReorderable}
\title{gtkNotebookGetTabReorderable}
\description{Gets whether the tab can be reordered via drag and drop or not.}
\usage{gtkNotebookGetTabReorderable(object, child)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkNotebook}}}
\item{\verb{child}}{a child \code{\link{GtkWidget}}}
}
\details{Since 2.10}
\value{[logical] \code{TRUE} if the tab is reorderable.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
