\alias{gtkNotebookNew}
\name{gtkNotebookNew}
\title{gtkNotebookNew}
\description{Creates a new \code{\link{GtkNotebook}} widget with no pages.}
\usage{gtkNotebookNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] the newly created \code{\link{GtkNotebook}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
