\alias{gtkNotebookNextPage}
\name{gtkNotebookNextPage}
\title{gtkNotebookNextPage}
\description{Switches to the next page. Nothing happens if the current page is
the last page.}
\usage{gtkNotebookNextPage(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkNotebook}}]  a \verb{\link{GtkNotebook}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
