\alias{gtkNotebookPopupDisable}
\name{gtkNotebookPopupDisable}
\title{gtkNotebookPopupDisable}
\description{Disables the popup menu.}
\usage{gtkNotebookPopupDisable(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkNotebook}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
