\alias{gtkNotebookPopupEnable}
\name{gtkNotebookPopupEnable}
\title{gtkNotebookPopupEnable}
\description{Enables the popup menu: if the user clicks with the right mouse button on
the tab labels, a menu with all the pages will be popped up.}
\usage{gtkNotebookPopupEnable(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkNotebook}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
