\alias{gtkNotebookPrevPage}
\name{gtkNotebookPrevPage}
\title{gtkNotebookPrevPage}
\description{Switches to the previous page. Nothing happens if the current page
is the first page.}
\usage{gtkNotebookPrevPage(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkNotebook}}]  a \code{\link{GtkNotebook}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
