\alias{gtkNotebookSetMenuLabel}
\name{gtkNotebookSetMenuLabel}
\title{gtkNotebookSetMenuLabel}
\description{Changes the menu label for the page containing \code{child}.}
\usage{gtkNotebookSetMenuLabel(object, child, menu.label = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkNotebook}}]  a \verb{\link{GtkNotebook}}}
\item{\verb{child}}{[\verb{\link{GtkWidget}}]  the child widget}
\item{\verb{menu.label}}{[\verb{\link{GtkWidget}}]  the menu label, or NULL for default}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
