\alias{gtkNotebookSetMenuLabel}
\name{gtkNotebookSetMenuLabel}
\title{gtkNotebookSetMenuLabel}
\description{Changes the menu label for the page containing \code{child}.}
\usage{gtkNotebookSetMenuLabel(object, child, menu.label = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkNotebook}}}
\item{\verb{child}}{the child widget}
\item{\verb{menu.label}}{the menu label, or NULL for default. \emph{[  \acronym{allow-none}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
