\alias{gtkNotebookSetMenuLabelText}
\name{gtkNotebookSetMenuLabelText}
\title{gtkNotebookSetMenuLabelText}
\description{Creates a new label and sets it as the menu label of \code{child}.}
\usage{gtkNotebookSetMenuLabelText(object, child, menu.text)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkNotebook}}]  a \verb{\link{GtkNotebook}}}
\item{\verb{child}}{[\verb{\link{GtkWidget}}]  the child widget}
\item{\verb{menu.text}}{[character]  the label text}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
