\alias{gtkNotebookSetShowBorder}
\name{gtkNotebookSetShowBorder}
\title{gtkNotebookSetShowBorder}
\description{Sets whether a bevel will be drawn around the notebook pages.
This only has a visual effect when the tabs are not shown.
See \code{\link{gtkNotebookSetShowTabs}}.}
\usage{gtkNotebookSetShowBorder(object, show.border)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkNotebook}}}
\item{\verb{show.border}}{\code{TRUE} if a bevel should be drawn around the notebook.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
