\alias{gtkNotebookSetShowTabs}
\name{gtkNotebookSetShowTabs}
\title{gtkNotebookSetShowTabs}
\description{Sets whether to show the tabs for the notebook or not.}
\usage{gtkNotebookSetShowTabs(object, show.tabs)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkNotebook}}}
\item{\verb{show.tabs}}{\code{TRUE} if the tabs should be shown.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
