\alias{gtkNotebookSetTabLabelText}
\name{gtkNotebookSetTabLabelText}
\title{gtkNotebookSetTabLabelText}
\description{Creates a new label and sets it as the tab label for the page
containing \code{child}.}
\usage{gtkNotebookSetTabLabelText(object, child, tab.text)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkNotebook}}]  a \verb{\link{GtkNotebook}}}
\item{\verb{child}}{[\verb{\link{GtkWidget}}]  the page}
\item{\verb{tab.text}}{[character]  the label text}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
