\alias{gtkNotebookSetTabLabelText}
\name{gtkNotebookSetTabLabelText}
\title{gtkNotebookSetTabLabelText}
\description{Creates a new label and sets it as the tab label for the page
containing \code{child}.}
\usage{gtkNotebookSetTabLabelText(object, child, tab.text)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkNotebook}}}
\item{\verb{child}}{the page}
\item{\verb{tab.text}}{the label text}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
