\alias{gtkNotebookSetTabPos}
\name{gtkNotebookSetTabPos}
\title{gtkNotebookSetTabPos}
\description{Sets the edge at which the tabs for switching pages in the
notebook are drawn.}
\usage{gtkNotebookSetTabPos(object, pos)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkNotebook}}.}
\item{\verb{pos}}{the edge to draw the tabs at.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
