\alias{gtkNotebookSetTabReorderable}
\name{gtkNotebookSetTabReorderable}
\title{gtkNotebookSetTabReorderable}
\description{Sets whether the notebook tab can be reordered
via drag and drop or not.}
\usage{gtkNotebookSetTabReorderable(object, child, reorderable)}
\arguments{
\item{\code{object}}{[\code{\link{GtkNotebook}}]  a \code{\link{GtkNotebook}}}
\item{\code{child}}{[\code{\link{GtkWidget}}]  a child \code{\link{GtkWidget}}}
\item{\code{reorderable}}{[logical]  whether the tab is reorderable or not.}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
