\alias{gtkNotebookSetTabReorderable}
\name{gtkNotebookSetTabReorderable}
\title{gtkNotebookSetTabReorderable}
\description{Sets whether the notebook tab can be reordered
via drag and drop or not.}
\usage{gtkNotebookSetTabReorderable(object, child, reorderable)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkNotebook}}}
\item{\verb{child}}{a child \code{\link{GtkWidget}}}
\item{\verb{reorderable}}{whether the tab is reorderable or not.}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
