\alias{gtkOptionMenuGetMenu}
\name{gtkOptionMenuGetMenu}
\title{gtkOptionMenuGetMenu}
\description{
Returns the \code{\link{GtkMenu}} associated with the \code{\link{GtkOptionMenu}}.
\strong{WARNING: \code{gtk_option_menu_get_menu} is deprecated and should not be used in newly-written code. Use \code{\link{GtkComboBox}} instead.  
Returns the \code{\link{GtkMenu}} associated with the \code{\link{GtkOptionMenu}}.}
}
\usage{gtkOptionMenuGetMenu(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkOptionMenu}}] a \code{\link{GtkOptionMenu}}.}}
\value{[\code{\link{GtkWidget}}] the \code{\link{GtkMenu}} associated with the \code{\link{GtkOptionMenu}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
