\alias{gtkOptionMenuGetMenu}
\name{gtkOptionMenuGetMenu}
\title{gtkOptionMenuGetMenu}
\description{
Returns the \code{\link{GtkMenu}} associated with the \code{\link{GtkOptionMenu}}.
\strong{WARNING: \code{gtk_option_menu_get_menu} has been deprecated since version 2.4 and should not be used in newly-written code. Use \code{\link{GtkComboBox}} instead.}
}
\usage{gtkOptionMenuGetMenu(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkOptionMenu}}.}}
\value{[\code{\link{GtkWidget}}] the \code{\link{GtkMenu}} associated with the \code{\link{GtkOptionMenu}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
