\alias{gtkOptionMenuNew}
\name{gtkOptionMenuNew}
\title{gtkOptionMenuNew}
\description{
Creates a new \code{\link{GtkOptionMenu}}.
\strong{WARNING: \code{gtk_option_menu_new} is deprecated and should not be used in newly-written code. Use \code{\link{GtkComboBox}} instead.  
Creates a new \code{\link{GtkOptionMenu}}.}
}
\usage{gtkOptionMenuNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkOptionMenu}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
