\alias{gtkPageSetupCopy}
\name{gtkPageSetupCopy}
\title{gtkPageSetupCopy}
\description{Copies a \code{\link{GtkPageSetup}}.}
\usage{gtkPageSetupCopy(object)}
\arguments{\item{\verb{object}}{the \code{\link{GtkPageSetup}} to copy}}
\details{Since 2.10}
\value{[\code{\link{GtkPageSetup}}] a copy of \code{other}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
