\alias{gtkPageSetupGetLeftMargin}
\name{gtkPageSetupGetLeftMargin}
\title{gtkPageSetupGetLeftMargin}
\description{Gets the left margin in units of \code{unit}.}
\usage{gtkPageSetupGetLeftMargin(object, unit)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPageSetup}}}
\item{\verb{unit}}{the unit for the return value}
}
\details{Since 2.10}
\value{[numeric] the left margin}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
