\alias{gtkPageSetupGetOrientation}
\name{gtkPageSetupGetOrientation}
\title{gtkPageSetupGetOrientation}
\description{Gets the page orientation of the \verb{\link{GtkPageSetup}}.}
\usage{gtkPageSetupGetOrientation(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkPageSetup}}]  a \verb{\link{GtkPageSetup}}}}
\details{  Since  2.10}
\value{[\code{\link{GtkPageOrientation}}]  the page orientation}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
