\alias{gtkPageSetupGetOrientation}
\name{gtkPageSetupGetOrientation}
\title{gtkPageSetupGetOrientation}
\description{Gets the page orientation of the \code{\link{GtkPageSetup}}.}
\usage{gtkPageSetupGetOrientation(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPageSetup}}}}
\details{Since 2.10}
\value{[\code{\link{GtkPageOrientation}}] the page orientation}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
