\alias{gtkPageSetupGetPaperSize}
\name{gtkPageSetupGetPaperSize}
\title{gtkPageSetupGetPaperSize}
\description{Gets the paper size of the \verb{\link{GtkPageSetup}}.}
\usage{gtkPageSetupGetPaperSize(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkPageSetup}}]  a \verb{\link{GtkPageSetup}}}}
\details{  Since  2.10}
\value{[\verb{\link{GtkPaperSize}}]  the paper size}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
