\alias{gtkPageSetupGetPaperSize}
\name{gtkPageSetupGetPaperSize}
\title{gtkPageSetupGetPaperSize}
\description{Gets the paper size of the \code{\link{GtkPageSetup}}.}
\usage{gtkPageSetupGetPaperSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPageSetup}}}}
\details{Since 2.10}
\value{[\code{\link{GtkPaperSize}}] the paper size}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
