\alias{gtkPageSetupNew}
\name{gtkPageSetupNew}
\title{gtkPageSetupNew}
\description{Creates a new \verb{\link{GtkPageSetup}}.}
\usage{gtkPageSetupNew()}
\details{  Since  2.10}
\value{[\verb{\link{GtkPageSetup}}]  a new \verb{\link{GtkPageSetup}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
