\alias{gtkPageSetupSetBottomMargin}
\name{gtkPageSetupSetBottomMargin}
\title{gtkPageSetupSetBottomMargin}
\description{Sets the bottom margin of the \code{\link{GtkPageSetup}}.}
\usage{gtkPageSetupSetBottomMargin(object, margin, unit)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPageSetup}}]  a \code{\link{GtkPageSetup}}}
\item{\code{margin}}{[numeric]  the new bottom margin in units of \code{unit}}
\item{\code{unit}}{[\code{\link{GtkUnit}}]  the units for \code{margin}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
