\alias{gtkPageSetupSetBottomMargin}
\name{gtkPageSetupSetBottomMargin}
\title{gtkPageSetupSetBottomMargin}
\description{Sets the bottom margin of the \code{\link{GtkPageSetup}}.}
\usage{gtkPageSetupSetBottomMargin(object, margin, unit)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPageSetup}}}
\item{\verb{margin}}{the new bottom margin in units of \code{unit}}
\item{\verb{unit}}{the units for \code{margin}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
