\alias{gtkPageSetupSetLeftMargin}
\name{gtkPageSetupSetLeftMargin}
\title{gtkPageSetupSetLeftMargin}
\description{Sets the left margin of the \code{\link{GtkPageSetup}}.}
\usage{gtkPageSetupSetLeftMargin(object, margin, unit)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPageSetup}}}
\item{\verb{margin}}{the new left margin in units of \code{unit}}
\item{\verb{unit}}{the units for \code{margin}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
