\alias{gtkPageSetupSetPaperSize}
\name{gtkPageSetupSetPaperSize}
\title{gtkPageSetupSetPaperSize}
\description{Sets the paper size of the \verb{\link{GtkPageSetup}} without
changing the margins. See 
\code{\link{gtkPageSetupSetPaperSizeAndDefaultMargins}}.}
\usage{gtkPageSetupSetPaperSize(object, size)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPageSetup}}]  a \verb{\link{GtkPageSetup}}}
\item{\verb{size}}{[\verb{\link{GtkPaperSize}}]  a \verb{\link{GtkPaperSize}} }
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
