\alias{gtkPageSetupSetPaperSize}
\name{gtkPageSetupSetPaperSize}
\title{gtkPageSetupSetPaperSize}
\description{Sets the paper size of the \code{\link{GtkPageSetup}} without
changing the margins. See 
\code{\link{gtkPageSetupSetPaperSizeAndDefaultMargins}}.}
\usage{gtkPageSetupSetPaperSize(object, size)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPageSetup}}}
\item{\verb{size}}{a \code{\link{GtkPaperSize}}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
