\alias{gtkPageSetupSetPaperSizeAndDefaultMargins}
\name{gtkPageSetupSetPaperSizeAndDefaultMargins}
\title{gtkPageSetupSetPaperSizeAndDefaultMargins}
\description{Sets the paper size of the \verb{\link{GtkPageSetup}} and modifies
the margins according to the new paper size.}
\usage{gtkPageSetupSetPaperSizeAndDefaultMargins(object, size)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPageSetup}}]  a \verb{\link{GtkPageSetup}}}
\item{\verb{size}}{[\verb{\link{GtkPaperSize}}]  a \verb{\link{GtkPaperSize}} }
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
