\alias{gtkPageSetupSetPaperSizeAndDefaultMargins}
\name{gtkPageSetupSetPaperSizeAndDefaultMargins}
\title{gtkPageSetupSetPaperSizeAndDefaultMargins}
\description{Sets the paper size of the \code{\link{GtkPageSetup}} and modifies
the margins according to the new paper size.}
\usage{gtkPageSetupSetPaperSizeAndDefaultMargins(object, size)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPageSetup}}}
\item{\verb{size}}{a \code{\link{GtkPaperSize}}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
