\alias{gtkPageSetupSetRightMargin}
\name{gtkPageSetupSetRightMargin}
\title{gtkPageSetupSetRightMargin}
\description{Sets the right margin of the \code{\link{GtkPageSetup}}.}
\usage{gtkPageSetupSetRightMargin(object, margin, unit)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPageSetup}}}
\item{\verb{margin}}{the new right margin in units of \code{unit}}
\item{\verb{unit}}{the units for \code{margin}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
