\alias{gtkPageSetupSetTopMargin}
\name{gtkPageSetupSetTopMargin}
\title{gtkPageSetupSetTopMargin}
\description{Sets the top margin of the \code{\link{GtkPageSetup}}.}
\usage{gtkPageSetupSetTopMargin(object, margin, unit)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPageSetup}}}
\item{\verb{margin}}{the new top margin in units of \code{unit}}
\item{\verb{unit}}{the units for \code{margin}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
