\alias{gtkPaintSlider}
\name{gtkPaintSlider}
\title{gtkPaintSlider}
\description{\emph{undocumented
}}
\usage{gtkPaintSlider(object, window, state.type, shadow.type, area = NULL, widget = NULL, detail = NULL, x, y, width, height, orientation)}
\arguments{
\item{\code{object}}{[\code{\link{GtkStyle}}] }
\item{\code{window}}{[\code{\link{GdkWindow}}] }
\item{\code{state.type}}{[\code{\link{GtkStateType}}] }
\item{\code{shadow.type}}{[\code{\link{GtkShadowType}}] }
\item{\code{area}}{[\code{\link{GdkRectangle}}] }
\item{\code{widget}}{[\code{\link{GtkWidget}}] }
\item{\code{detail}}{[character] }
\item{\code{x}}{[integer] }
\item{\code{y}}{[integer] }
\item{\code{width}}{[integer] }
\item{\code{height}}{[integer] }
\item{\code{orientation}}{[\code{\link{GtkOrientation}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
