\alias{gtkPanedAdd2}
\name{gtkPanedAdd2}
\title{gtkPanedAdd2}
\description{Adds a child to the bottom or right pane with default
parameters. This is equivalent to
\code{gtk_paned_pack2 (paned, child, TRUE, TRUE)}.}
\usage{gtkPanedAdd2(object, child)}
\arguments{
\item{\verb{object}}{a paned widget}
\item{\verb{child}}{the child to add}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
