\alias{gtkPanedPack1}
\name{gtkPanedPack1}
\title{gtkPanedPack1}
\description{Adds a child to the top or left pane.}
\usage{gtkPanedPack1(object, child, resize = FALSE, shrink = TRUE)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPaned}}] a paned widget}
\item{\verb{child}}{[\verb{\link{GtkWidget}}] the child to add}
\item{\verb{resize}}{[logical] should this child expand when the paned widget is resized.}
\item{\verb{shrink}}{[logical] can this child be made smaller than its requisition.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
