\alias{gtkPanedPack1}
\name{gtkPanedPack1}
\title{gtkPanedPack1}
\description{Adds a child to the top or left pane.}
\usage{gtkPanedPack1(object, child, resize = FALSE, shrink = TRUE)}
\arguments{
\item{\verb{object}}{a paned widget}
\item{\verb{child}}{the child to add}
\item{\verb{resize}}{should this child expand when the paned widget is resized.}
\item{\verb{shrink}}{can this child be made smaller than its requisition.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
