\alias{gtkPanedPack2}
\name{gtkPanedPack2}
\title{gtkPanedPack2}
\description{Adds a child to the bottom or right pane.}
\usage{gtkPanedPack2(object, child, resize = TRUE, shrink = TRUE)}
\arguments{
\item{\verb{object}}{a paned widget}
\item{\verb{child}}{the child to add}
\item{\verb{resize}}{should this child expand when the paned widget is resized.}
\item{\verb{shrink}}{can this child be made smaller than its requisition.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
