\alias{gtkPanedSetPosition}
\name{gtkPanedSetPosition}
\title{gtkPanedSetPosition}
\description{Sets the position of the divider between the two panes.}
\usage{gtkPanedSetPosition(object, position)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPaned}} widget}
\item{\verb{position}}{pixel position of divider, a negative value means that the position
is unset.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
