\alias{gtkPaperSizeCopy}
\name{gtkPaperSizeCopy}
\title{gtkPaperSizeCopy}
\description{Copies an existing \code{\link{GtkPaperSize}}.}
\usage{gtkPaperSizeCopy(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPaperSize}}}}
\details{Since 2.10}
\value{[\code{\link{GtkPaperSize}}] a copy of \code{other}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
