\alias{gtkPaperSizeGetDefaultLeftMargin}
\name{gtkPaperSizeGetDefaultLeftMargin}
\title{gtkPaperSizeGetDefaultLeftMargin}
\description{Gets the default left margin for the \code{\link{GtkPaperSize}}.}
\usage{gtkPaperSizeGetDefaultLeftMargin(object, unit)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPaperSize}} object}
\item{\verb{unit}}{the unit for the return value}
}
\details{Since 2.10}
\value{[numeric] the default left margin}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
