\alias{gtkPaperSizeGetDefaultRightMargin}
\name{gtkPaperSizeGetDefaultRightMargin}
\title{gtkPaperSizeGetDefaultRightMargin}
\description{Gets the default right margin for the \code{\link{GtkPaperSize}}.}
\usage{gtkPaperSizeGetDefaultRightMargin(object, unit)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPaperSize}} object}
\item{\verb{unit}}{the unit for the return value}
}
\details{Since 2.10}
\value{[numeric] the default right margin}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
