\alias{gtkPaperSizeGetDefaultTopMargin}
\name{gtkPaperSizeGetDefaultTopMargin}
\title{gtkPaperSizeGetDefaultTopMargin}
\description{Gets the default top margin for the \code{\link{GtkPaperSize}}.}
\usage{gtkPaperSizeGetDefaultTopMargin(object, unit)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPaperSize}} object}
\item{\verb{unit}}{the unit for the return value}
}
\details{Since 2.10}
\value{[numeric] the default top margin}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
