\alias{gtkPaperSizeIsEqual}
\name{gtkPaperSizeIsEqual}
\title{gtkPaperSizeIsEqual}
\description{Compares two \code{\link{GtkPaperSize}} objects.}
\usage{gtkPaperSizeIsEqual(object, size2)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPaperSize}}]  a \code{\link{GtkPaperSize}} object}
\item{\code{size2}}{[\code{\link{GtkPaperSize}}]  another \code{\link{GtkPaperSize}} object}
}
\details{  Since  2.10}
\value{[logical] \code{TRUE}, if \code{size1} and \code{size2} 
represent the same paper size}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
