\alias{gtkPaperSizeNewCustom}
\name{gtkPaperSizeNewCustom}
\title{gtkPaperSizeNewCustom}
\description{Creates a new \code{\link{GtkPaperSize}} object with the
given parameters.}
\usage{gtkPaperSizeNewCustom(name, display.name, width, height, unit)}
\arguments{
\item{\verb{name}}{the paper name}
\item{\verb{display.name}}{the human-readable name}
\item{\verb{width}}{the paper width, in units of \code{unit}}
\item{\verb{height}}{the paper height, in units of \code{unit}}
\item{\verb{unit}}{the unit for \code{width} and \code{height}}
}
\details{Since 2.10}
\value{[\code{\link{GtkPaperSize}}] a new \code{\link{GtkPaperSize}} object,}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
