\alias{gtkPaperSizeSetSize}
\name{gtkPaperSizeSetSize}
\title{gtkPaperSizeSetSize}
\description{Changes the dimensions of a \code{size} to \code{width} x \code{height}.}
\usage{gtkPaperSizeSetSize(object, width, height, unit)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPaperSize}}]  a custom \code{\link{GtkPaperSize}} object}
\item{\code{width}}{[numeric]  the new width in units of \code{unit}}
\item{\code{height}}{[numeric]  the new height in units of \code{unit}}
\item{\code{unit}}{[\code{\link{GtkUnit}}]  the unit for \code{width} and \code{height}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
