\alias{gtkPaperSizeSetSize}
\name{gtkPaperSizeSetSize}
\title{gtkPaperSizeSetSize}
\description{Changes the dimensions of a \code{size} to \code{width} x \code{height}.}
\usage{gtkPaperSizeSetSize(object, width, height, unit)}
\arguments{
\item{\verb{object}}{a custom \code{\link{GtkPaperSize}} object}
\item{\verb{width}}{the new width in units of \code{unit}}
\item{\verb{height}}{the new height in units of \code{unit}}
\item{\verb{unit}}{the unit for \code{width} and \code{height}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
