\alias{gtkPaperSizeToKeyFile}
\name{gtkPaperSizeToKeyFile}
\title{gtkPaperSizeToKeyFile}
\description{This function adds the paper size from \code{size} to \code{key.file}.}
\usage{gtkPaperSizeToKeyFile(object, key.file, group.name)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPaperSize}}]  a \code{\link{GtkPaperSize}}}
\item{\code{key.file}}{[GKeyFile]  the \code{GKeyFile} to save the paper size to}
\item{\code{group.name}}{[character]  the group to add the settings to in \code{key.file}}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
