\alias{gtkPaperSizeToKeyFile}
\name{gtkPaperSizeToKeyFile}
\title{gtkPaperSizeToKeyFile}
\description{This function adds the paper size from \code{size} to \code{key.file}.}
\usage{gtkPaperSizeToKeyFile(object, key.file, group.name)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPaperSize}}}
\item{\verb{key.file}}{the \verb{GKeyFile} to save the paper size to}
\item{\verb{group.name}}{the group to add the settings to in \code{key.file}}
}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
