\alias{gtkPlugGetId}
\name{gtkPlugGetId}
\title{gtkPlugGetId}
\description{Gets the window ID of a \code{\link{GtkPlug}} widget, which can then
be used to embed this window inside another window, for
instance with \code{\link{gtkSocketAddId}}.}
\usage{gtkPlugGetId(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPlug}}.}}
\value{[\code{\link{GdkNativeWindow}}] the window ID for the plug}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
