\alias{gtkPreviewGetInfo}
\name{gtkPreviewGetInfo}
\title{gtkPreviewGetInfo}
\description{
Return a \code{\link{GtkPreviewInfo}} structure containing
global information about preview widgets.
\strong{WARNING: \code{gtk_preview_get_info} is deprecated and should not be used in newly-written code.}
}
\usage{gtkPreviewGetInfo()}
\value{[\code{\link{GtkPreviewInfo}}] a \code{\link{GtkPreviewInfo}} structure. The return
value belongs to GTK+ and must not be modified
or freed.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
