\alias{gtkPreviewNew}
\name{gtkPreviewNew}
\title{gtkPreviewNew}
\description{
Create a new preview widget.
\strong{WARNING: \code{gtk_preview_new} is deprecated and should not be used in newly-written code.}
}
\usage{gtkPreviewNew(type, show = TRUE)}
\arguments{\item{\verb{type}}{the type data contained by the widget.
(Grayscale or RGB)}}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkPreview}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
