\alias{gtkPreviewSetInstallCmap}
\name{gtkPreviewSetInstallCmap}
\title{gtkPreviewSetInstallCmap}
\description{
This function is deprecated
and does nothing. GdkRGB will automatically pick
a private colormap if it cannot allocate sufficient
colors.
\strong{WARNING: \code{gtk_preview_set_install_cmap} is deprecated and should not be used in newly-written code.}
}
\usage{gtkPreviewSetInstallCmap(install.cmap)}
\arguments{\item{\verb{install.cmap}}{ignored.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
