\alias{gtkPreviewSetReserved}
\name{gtkPreviewSetReserved}
\title{gtkPreviewSetReserved}
\description{
This function is deprecated and does nothing.
\strong{WARNING: \code{gtk_preview_set_reserved} is deprecated and should not be used in newly-written code.}
}
\usage{gtkPreviewSetReserved(nreserved)}
\arguments{\item{\verb{nreserved}}{ignored.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
