\alias{gtkPrintContextCreatePangoContext}
\name{gtkPrintContextCreatePangoContext}
\title{gtkPrintContextCreatePangoContext}
\description{Creates a new \code{\link{PangoContext}} that can be used with the
\code{\link{GtkPrintContext}}.}
\usage{gtkPrintContextCreatePangoContext(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintContext}}}}
\details{Since 2.10}
\value{[\code{\link{PangoContext}}] a new Pango context for \code{context}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
