\alias{gtkPrintContextCreatePangoLayout}
\name{gtkPrintContextCreatePangoLayout}
\title{gtkPrintContextCreatePangoLayout}
\description{Creates a new \code{\link{PangoLayout}} that is suitable for use
with the \code{\link{GtkPrintContext}}.}
\usage{gtkPrintContextCreatePangoLayout(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintContext}}}}
\details{Since 2.10}
\value{[\code{\link{PangoLayout}}] a new Pango layout for \code{context}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
