\alias{gtkPrintContextGetCairoContext}
\name{gtkPrintContextGetCairoContext}
\title{gtkPrintContextGetCairoContext}
\description{Obtains the cairo context that is associated with the
\code{\link{GtkPrintContext}}.}
\usage{gtkPrintContextGetCairoContext(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintContext}}}}
\details{Since 2.10}
\value{[\code{\link{Cairo}}] the cairo context of \code{context}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
