\alias{gtkPrintContextGetDpiX}
\name{gtkPrintContextGetDpiX}
\title{gtkPrintContextGetDpiX}
\description{Obtains the horizontal resolution of the \code{\link{GtkPrintContext}},
in dots per inch.}
\usage{gtkPrintContextGetDpiX(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintContext}}}}
\details{Since 2.10}
\value{[numeric] the horizontal resolution of \code{context}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
