\alias{gtkPrintContextGetDpiY}
\name{gtkPrintContextGetDpiY}
\title{gtkPrintContextGetDpiY}
\description{Obtains the vertical resolution of the \code{\link{GtkPrintContext}},
in dots per inch.}
\usage{gtkPrintContextGetDpiY(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkPrintContext}}]  a \code{\link{GtkPrintContext}}}}
\details{  Since  2.10}
\value{[numeric]  the vertical resolution of \code{context}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
