\alias{gtkPrintContextGetHeight}
\name{gtkPrintContextGetHeight}
\title{gtkPrintContextGetHeight}
\description{Obtains the height of the \code{\link{GtkPrintContext}}, in pixels.}
\usage{gtkPrintContextGetHeight(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintContext}}}}
\details{Since 2.10}
\value{[numeric] the height of \code{context}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
