\alias{gtkPrintContextGetPageSetup}
\name{gtkPrintContextGetPageSetup}
\title{gtkPrintContextGetPageSetup}
\description{Obtains the \code{\link{GtkPageSetup}} that determines the page
dimensions of the \code{\link{GtkPrintContext}}.}
\usage{gtkPrintContextGetPageSetup(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintContext}}}}
\details{Since 2.10}
\value{[\code{\link{GtkPageSetup}}] the page setup of \code{context}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
