\alias{gtkPrintContextGetPangoFontmap}
\name{gtkPrintContextGetPangoFontmap}
\title{gtkPrintContextGetPangoFontmap}
\description{Returns a \code{\link{PangoFontMap}} that is suitable for use 
with the \code{\link{GtkPrintContext}}.}
\usage{gtkPrintContextGetPangoFontmap(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintContext}}}}
\details{Since 2.10}
\value{[\code{\link{PangoFontMap}}] the font map of \code{context}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
