\alias{gtkPrintOperationGetDefaultPageSetup}
\name{gtkPrintOperationGetDefaultPageSetup}
\title{gtkPrintOperationGetDefaultPageSetup}
\description{Returns the default page setup, see 
\code{\link{gtkPrintOperationSetDefaultPageSetup}}.}
\usage{gtkPrintOperationGetDefaultPageSetup(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintOperation}}}}
\details{Since 2.10}
\value{[\code{\link{GtkPageSetup}}] the default page setup}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
