\alias{gtkPrintOperationNew}
\name{gtkPrintOperationNew}
\title{gtkPrintOperationNew}
\description{Creates a new \code{\link{GtkPrintOperation}}.}
\usage{gtkPrintOperationNew()}
\details{Since 2.10}
\value{[\code{\link{GtkPrintOperation}}] a new \code{\link{GtkPrintOperation}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
