\alias{gtkPrintOperationSetCustomTabLabel}
\name{gtkPrintOperationSetCustomTabLabel}
\title{gtkPrintOperationSetCustomTabLabel}
\description{Sets the label for the tab holding custom widgets.}
\usage{gtkPrintOperationSetCustomTabLabel(object, label)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintOperation}}}
\item{\verb{label}}{the label to use, or \code{NULL} to use the default label. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
