\alias{gtkPrintOperationSetPrintSettings}
\name{gtkPrintOperationSetPrintSettings}
\title{gtkPrintOperationSetPrintSettings}
\description{Sets the print settings for \code{op}. This is typically used to
re-establish print settings from a previous print operation,
see \code{\link{gtkPrintOperationRun}}.}
\usage{gtkPrintOperationSetPrintSettings(object, print.settings = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintOperation}}]  a \code{\link{GtkPrintOperation}}}
\item{\code{print.settings}}{[\code{\link{GtkPrintSettings}}] \code{\link{GtkPrintSettings}}, or \code{NULL}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
