\alias{gtkPrintOperationSetShowProgress}
\name{gtkPrintOperationSetShowProgress}
\title{gtkPrintOperationSetShowProgress}
\description{If \code{show.progress} is \code{TRUE}, the print operation will show a 
progress dialog during the print operation.}
\usage{gtkPrintOperationSetShowProgress(object, show.progress)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintOperation}}]  a \code{\link{GtkPrintOperation}}}
\item{\code{show.progress}}{[logical] \code{TRUE} to show a progress dialog}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
