\alias{gtkPrintOperationSetShowProgress}
\name{gtkPrintOperationSetShowProgress}
\title{gtkPrintOperationSetShowProgress}
\description{If \code{show.progress} is \code{TRUE}, the print operation will show a 
progress dialog during the print operation.}
\usage{gtkPrintOperationSetShowProgress(object, show.progress)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintOperation}}}
\item{\verb{show.progress}}{\code{TRUE} to show a progress dialog}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
