\alias{gtkPrintOperationSetUnit}
\name{gtkPrintOperationSetUnit}
\title{gtkPrintOperationSetUnit}
\description{Sets up the transformation for the cairo context obtained from
\verb{\link{GtkPrintContext}} in such a way that distances are measured in 
units of \code{unit}.}
\usage{gtkPrintOperationSetUnit(object, unit)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPrintOperation}}]  a \verb{\link{GtkPrintOperation}}}
\item{\verb{unit}}{[\code{\link{GtkUnit}}]  the unit to use}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
