\alias{gtkPrintSettingsCopy}
\name{gtkPrintSettingsCopy}
\title{gtkPrintSettingsCopy}
\description{Copies a \code{\link{GtkPrintSettings}} object.}
\usage{gtkPrintSettingsCopy(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.10}
\value{[\code{\link{GtkPrintSettings}}] a newly allocated copy of \code{other}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
