\alias{gtkPrintSettingsForeach}
\name{gtkPrintSettingsForeach}
\title{gtkPrintSettingsForeach}
\description{Calls \code{func} for each key-value pair of \code{settings}.}
\usage{gtkPrintSettingsForeach(object, func, user.data = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{func}}{[\code{\link{GtkPrintSettingsFunc}}]  the function to call}
\item{\code{user.data}}{[R object]  user data for \code{func}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
