\alias{gtkPrintSettingsForeach}
\name{gtkPrintSettingsForeach}
\title{gtkPrintSettingsForeach}
\description{Calls \code{func} for each key-value pair of \code{settings}.}
\usage{gtkPrintSettingsForeach(object, func, user.data = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{func}}{(scope call) the function to call}
\item{\verb{user.data}}{user data for \code{func}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
