\alias{gtkPrintSettingsGet}
\name{gtkPrintSettingsGet}
\title{gtkPrintSettingsGet}
\description{Looks up the string value associated with \code{key}.}
\usage{gtkPrintSettingsGet(object, key)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{key}}{a key}
}
\details{Since 2.10}
\value{[character] the string value for \code{key}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
