\alias{gtkPrintSettingsGetDouble}
\name{gtkPrintSettingsGetDouble}
\title{gtkPrintSettingsGetDouble}
\description{Returns the double value associated with \code{key}, or 0.}
\usage{gtkPrintSettingsGetDouble(object, key)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{key}}{[character]  a key}
}
\details{  Since  2.10}
\value{[numeric]  the double value of \code{key}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
