\alias{gtkPrintSettingsGetDouble}
\name{gtkPrintSettingsGetDouble}
\title{gtkPrintSettingsGetDouble}
\description{Returns the double value associated with \code{key}, or 0.}
\usage{gtkPrintSettingsGetDouble(object, key)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{key}}{a key}
}
\details{Since 2.10}
\value{[numeric] the double value of \code{key}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
