\alias{gtkPrintSettingsGetFinishings}
\name{gtkPrintSettingsGetFinishings}
\title{gtkPrintSettingsGetFinishings}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_FINISHINGS}.}
\usage{gtkPrintSettingsGetFinishings(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.10}
\value{[character] the finishings}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
