\alias{gtkPrintSettingsGetInt}
\name{gtkPrintSettingsGetInt}
\title{gtkPrintSettingsGetInt}
\description{Returns the integer value of \code{key}, or 0.}
\usage{gtkPrintSettingsGetInt(object, key)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{key}}{a key}
}
\details{Since 2.10}
\value{[integer] the integer value of \code{key}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
