\alias{gtkPrintSettingsGetIntWithDefault}
\name{gtkPrintSettingsGetIntWithDefault}
\title{gtkPrintSettingsGetIntWithDefault}
\description{Returns the value of \code{key}, interpreted as
an integer, or the default value.}
\usage{gtkPrintSettingsGetIntWithDefault(object, key, def)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{key}}{a key}
\item{\verb{def}}{the default value}
}
\details{Since 2.10}
\value{[integer] the integer value of \code{key}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
