\alias{gtkPrintSettingsGetLength}
\name{gtkPrintSettingsGetLength}
\title{gtkPrintSettingsGetLength}
\description{Returns the value associated with \code{key}, interpreted
as a length. The returned value is converted to \code{units}.}
\usage{gtkPrintSettingsGetLength(object, key, unit)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{key}}{a key}
\item{\verb{unit}}{the unit of the return value}
}
\details{Since 2.10}
\value{[numeric] the length value of \code{key}, converted to \code{unit}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
