\alias{gtkPrintSettingsGetOrientation}
\name{gtkPrintSettingsGetOrientation}
\title{gtkPrintSettingsGetOrientation}
\description{Get the value of \code{GTK_PRINT_SETTINGS_ORIENTATION}, 
converted to a \code{\link{GtkPageOrientation}}.}
\usage{gtkPrintSettingsGetOrientation(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.10}
\value{[\code{\link{GtkPageOrientation}}] the orientation}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
