\alias{gtkPrintSettingsGetPageRanges}
\name{gtkPrintSettingsGetPageRanges}
\title{gtkPrintSettingsGetPageRanges}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_PAGE_RANGES}.}
\usage{gtkPrintSettingsGetPageRanges(object, num.ranges)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPrintSettings}}]  a \verb{\link{GtkPrintSettings}}}
\item{\verb{num.ranges}}{[integer]  return location for the length of the returned list}
}
\details{  Since  2.10}
\value{[\verb{\link{GtkPageRange}}]  a list of \verb{\link{GtkPageRange}}s. }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
