\alias{gtkPrintSettingsGetPaperSize}
\name{gtkPrintSettingsGetPaperSize}
\title{gtkPrintSettingsGetPaperSize}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_PAPER_FORMAT}, 
converted to a \code{\link{GtkPaperSize}}.}
\usage{gtkPrintSettingsGetPaperSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.10}
\value{[\code{\link{GtkPaperSize}}] the paper size}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
