\alias{gtkPrintSettingsGetQuality}
\name{gtkPrintSettingsGetQuality}
\title{gtkPrintSettingsGetQuality}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_QUALITY}.}
\usage{gtkPrintSettingsGetQuality(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.10}
\value{[\code{\link{GtkPrintQuality}}] the print quality}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
