\alias{gtkPrintSettingsGetUseColor}
\name{gtkPrintSettingsGetUseColor}
\title{gtkPrintSettingsGetUseColor}
\description{Gets the value of \code{GTK_PRINT_SETTINGS_USE_COLOR}.}
\usage{gtkPrintSettingsGetUseColor(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}}
\details{Since 2.10}
\value{[logical] whether to use color}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
