\alias{gtkPrintSettingsHasKey}
\name{gtkPrintSettingsHasKey}
\title{gtkPrintSettingsHasKey}
\description{Returns \code{TRUE}, if a value is associated with \code{key}.}
\usage{gtkPrintSettingsHasKey(object, key)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{key}}{[character]  a key}
}
\details{  Since  2.10}
\value{[logical] \code{TRUE}, if \code{key} has a value}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
