\alias{gtkPrintSettingsHasKey}
\name{gtkPrintSettingsHasKey}
\title{gtkPrintSettingsHasKey}
\description{Returns \code{TRUE}, if a value is associated with \code{key}.}
\usage{gtkPrintSettingsHasKey(object, key)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{key}}{a key}
}
\details{Since 2.10}
\value{[logical] \code{TRUE}, if \code{key} has a value}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
