\alias{gtkPrintSettingsNew}
\name{gtkPrintSettingsNew}
\title{gtkPrintSettingsNew}
\description{Creates a new \code{\link{GtkPrintSettings}} object.}
\usage{gtkPrintSettingsNew()}
\details{Since 2.10}
\value{[\code{\link{GtkPrintSettings}}] a new \code{\link{GtkPrintSettings}} object}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
