\alias{gtkPrintSettingsSet}
\name{gtkPrintSettingsSet}
\title{gtkPrintSettingsSet}
\description{Associates \code{value} with \code{key}.}
\usage{gtkPrintSettingsSet(object, key, value)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPrintSettings}}]  a \verb{\link{GtkPrintSettings}}}
\item{\verb{key}}{[character]  a key}
\item{\verb{value}}{[character]  a string value, or \code{NULL}}
}
\details{  Since  2.10 }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
