\alias{gtkPrintSettingsSet}
\name{gtkPrintSettingsSet}
\title{gtkPrintSettingsSet}
\description{Associates \code{value} with \code{key}.}
\usage{gtkPrintSettingsSet(object, key, value)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{key}}{a key}
\item{\verb{value}}{a string value, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
