\alias{gtkPrintSettingsSetBool}
\name{gtkPrintSettingsSetBool}
\title{gtkPrintSettingsSetBool}
\description{Sets \code{key} to a boolean value.}
\usage{gtkPrintSettingsSetBool(object, key, value)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPrintSettings}}]  a \verb{\link{GtkPrintSettings}}}
\item{\verb{key}}{[character]  a key}
\item{\verb{value}}{[logical]  a boolean}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
