\alias{gtkPrintSettingsSetCollate}
\name{gtkPrintSettingsSetCollate}
\title{gtkPrintSettingsSetCollate}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_COLLATE}.}
\usage{gtkPrintSettingsSetCollate(object, collate)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{collate}}{whether to collate the output}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
