\alias{gtkPrintSettingsSetDefaultSource}
\name{gtkPrintSettingsSetDefaultSource}
\title{gtkPrintSettingsSetDefaultSource}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_DEFAULT_SOURCE}.}
\usage{gtkPrintSettingsSetDefaultSource(object, default.source)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{default.source}}{the default source}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
