\alias{gtkPrintSettingsSetDouble}
\name{gtkPrintSettingsSetDouble}
\title{gtkPrintSettingsSetDouble}
\description{Sets \code{key} to a double value.}
\usage{gtkPrintSettingsSetDouble(object, key, value)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{key}}{[character]  a key }
\item{\code{value}}{[numeric]  a double value}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
