\alias{gtkPrintSettingsSetDouble}
\name{gtkPrintSettingsSetDouble}
\title{gtkPrintSettingsSetDouble}
\description{Sets \code{key} to a double value.}
\usage{gtkPrintSettingsSetDouble(object, key, value)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{key}}{a key}
\item{\verb{value}}{a double value}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
