\alias{gtkPrintSettingsSetDuplex}
\name{gtkPrintSettingsSetDuplex}
\title{gtkPrintSettingsSetDuplex}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_DUPLEX}.}
\usage{gtkPrintSettingsSetDuplex(object, duplex)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{duplex}}{[\code{\link{GtkPrintDuplex}}]  a \code{\link{GtkPrintDuplex}} value}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
