\alias{gtkPrintSettingsSetDuplex}
\name{gtkPrintSettingsSetDuplex}
\title{gtkPrintSettingsSetDuplex}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_DUPLEX}.}
\usage{gtkPrintSettingsSetDuplex(object, duplex)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{duplex}}{a \code{\link{GtkPrintDuplex}} value}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
