\alias{gtkPrintSettingsSetFinishings}
\name{gtkPrintSettingsSetFinishings}
\title{gtkPrintSettingsSetFinishings}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_FINISHINGS}.}
\usage{gtkPrintSettingsSetFinishings(object, finishings)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPrintSettings}}]  a \verb{\link{GtkPrintSettings}}}
\item{\verb{finishings}}{[character]  the finishings}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
