\alias{gtkPrintSettingsSetFinishings}
\name{gtkPrintSettingsSetFinishings}
\title{gtkPrintSettingsSetFinishings}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_FINISHINGS}.}
\usage{gtkPrintSettingsSetFinishings(object, finishings)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{finishings}}{the finishings}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
