\alias{gtkPrintSettingsSetInt}
\name{gtkPrintSettingsSetInt}
\title{gtkPrintSettingsSetInt}
\description{Sets \code{key} to an integer value.}
\usage{gtkPrintSettingsSetInt(object, key, value)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPrintSettings}}]  a \verb{\link{GtkPrintSettings}}}
\item{\verb{key}}{[character]  a key}
\item{\verb{value}}{[integer]  an integer }
}
\details{  Since  2.10 }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
