\alias{gtkPrintSettingsSetInt}
\name{gtkPrintSettingsSetInt}
\title{gtkPrintSettingsSetInt}
\description{Sets \code{key} to an integer value.}
\usage{gtkPrintSettingsSetInt(object, key, value)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{key}}{a key}
\item{\verb{value}}{an integer}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
