\alias{gtkPrintSettingsSetLength}
\name{gtkPrintSettingsSetLength}
\title{gtkPrintSettingsSetLength}
\description{Associates a length in units of \code{unit} with \code{key}.}
\usage{gtkPrintSettingsSetLength(object, key, value, unit)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{key}}{a key}
\item{\verb{value}}{a length}
\item{\verb{unit}}{the unit of \code{length}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
