\alias{gtkPrintSettingsSetOrientation}
\name{gtkPrintSettingsSetOrientation}
\title{gtkPrintSettingsSetOrientation}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_ORIENTATION}.}
\usage{gtkPrintSettingsSetOrientation(object, orientation)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{orientation}}{a page orientation}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
