\alias{gtkPrintSettingsSetOutputBin}
\name{gtkPrintSettingsSetOutputBin}
\title{gtkPrintSettingsSetOutputBin}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_OUTPUT_BIN}.}
\usage{gtkPrintSettingsSetOutputBin(object, output.bin)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{output.bin}}{the output bin}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
