\alias{gtkPrintSettingsSetPageRanges}
\name{gtkPrintSettingsSetPageRanges}
\title{gtkPrintSettingsSetPageRanges}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_PAGE_RANGES}.}
\usage{gtkPrintSettingsSetPageRanges(object, page.ranges, num.ranges)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{page.ranges}}{a list of \code{\link{GtkPageRange}}s}
\item{\verb{num.ranges}}{the length of \code{page.ranges}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
