\alias{gtkPrintSettingsSetPageSet}
\name{gtkPrintSettingsSetPageSet}
\title{gtkPrintSettingsSetPageSet}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_PAGE_SET}.}
\usage{gtkPrintSettingsSetPageSet(object, page.set)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{page.set}}{[\code{\link{GtkPageSet}}]  a \code{\link{GtkPageSet}} value}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
