\alias{gtkPrintSettingsSetPageSet}
\name{gtkPrintSettingsSetPageSet}
\title{gtkPrintSettingsSetPageSet}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_PAGE_SET}.}
\usage{gtkPrintSettingsSetPageSet(object, page.set)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{page.set}}{a \code{\link{GtkPageSet}} value}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
