\alias{gtkPrintSettingsSetPaperHeight}
\name{gtkPrintSettingsSetPaperHeight}
\title{gtkPrintSettingsSetPaperHeight}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_PAPER_HEIGHT}.}
\usage{gtkPrintSettingsSetPaperHeight(object, height, unit)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{height}}{[numeric]  the paper height}
\item{\code{unit}}{[\code{\link{GtkUnit}}]  the units of \code{height}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
