\alias{gtkPrintSettingsSetPaperHeight}
\name{gtkPrintSettingsSetPaperHeight}
\title{gtkPrintSettingsSetPaperHeight}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_PAPER_HEIGHT}.}
\usage{gtkPrintSettingsSetPaperHeight(object, height, unit)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{height}}{the paper height}
\item{\verb{unit}}{the units of \code{height}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
