\alias{gtkPrintSettingsSetPaperSize}
\name{gtkPrintSettingsSetPaperSize}
\title{gtkPrintSettingsSetPaperSize}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_PAPER_FORMAT},
\code{GTK_PRINT_SETTINGS_PAPER_WIDTH} and
\code{GTK_PRINT_SETTINGS_PAPER_HEIGHT}.}
\usage{gtkPrintSettingsSetPaperSize(object, paper.size)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPrintSettings}}]  a \verb{\link{GtkPrintSettings}}}
\item{\verb{paper.size}}{[\verb{\link{GtkPaperSize}}]  a paper size}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
