\alias{gtkPrintSettingsSetPaperWidth}
\name{gtkPrintSettingsSetPaperWidth}
\title{gtkPrintSettingsSetPaperWidth}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_PAPER_WIDTH}.}
\usage{gtkPrintSettingsSetPaperWidth(object, width, unit)}
\arguments{
\item{\code{object}}{[\code{\link{GtkPrintSettings}}]  a \code{\link{GtkPrintSettings}}}
\item{\code{width}}{[numeric]  the paper width}
\item{\code{unit}}{[\code{\link{GtkUnit}}]  the units of \code{width}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
