\alias{gtkPrintSettingsSetPaperWidth}
\name{gtkPrintSettingsSetPaperWidth}
\title{gtkPrintSettingsSetPaperWidth}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_PAPER_WIDTH}.}
\usage{gtkPrintSettingsSetPaperWidth(object, width, unit)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{width}}{the paper width}
\item{\verb{unit}}{the units of \code{width}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
