\alias{gtkPrintSettingsSetPrinter}
\name{gtkPrintSettingsSetPrinter}
\title{gtkPrintSettingsSetPrinter}
\description{Convenience function to set \code{GTK_PRINT_SETTINGS_PRINTER}
to \code{printer}.}
\usage{gtkPrintSettingsSetPrinter(object, printer)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPrintSettings}}]  a \verb{\link{GtkPrintSettings}}}
\item{\verb{printer}}{[character]  the printer name}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
