\alias{gtkPrintSettingsSetPrinter}
\name{gtkPrintSettingsSetPrinter}
\title{gtkPrintSettingsSetPrinter}
\description{Convenience function to set \code{GTK_PRINT_SETTINGS_PRINTER}
to \code{printer}.}
\usage{gtkPrintSettingsSetPrinter(object, printer)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{printer}}{the printer name}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
