\alias{gtkPrintSettingsSetPrintPages}
\name{gtkPrintSettingsSetPrintPages}
\title{gtkPrintSettingsSetPrintPages}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_PRINT_PAGES}.}
\usage{gtkPrintSettingsSetPrintPages(object, pages)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPrintSettings}}]  a \verb{\link{GtkPrintSettings}}}
\item{\verb{pages}}{[\code{\link{GtkPrintPages}}]  a \code{\link{GtkPrintPages}} value}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
