\alias{gtkPrintSettingsSetPrintPages}
\name{gtkPrintSettingsSetPrintPages}
\title{gtkPrintSettingsSetPrintPages}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_PRINT_PAGES}.}
\usage{gtkPrintSettingsSetPrintPages(object, pages)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{pages}}{a \code{\link{GtkPrintPages}} value}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
