\alias{gtkPrintSettingsSetQuality}
\name{gtkPrintSettingsSetQuality}
\title{gtkPrintSettingsSetQuality}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_QUALITY}.}
\usage{gtkPrintSettingsSetQuality(object, quality)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{quality}}{a \code{\link{GtkPrintQuality}} value}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
