\alias{gtkPrintSettingsSetResolution}
\name{gtkPrintSettingsSetResolution}
\title{gtkPrintSettingsSetResolution}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_RESOLUTION}.}
\usage{gtkPrintSettingsSetResolution(object, resolution)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkPrintSettings}}]  a \verb{\link{GtkPrintSettings}}}
\item{\verb{resolution}}{[integer]  the resolution in dpi}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
