\alias{gtkPrintSettingsSetResolution}
\name{gtkPrintSettingsSetResolution}
\title{gtkPrintSettingsSetResolution}
\description{Sets the values of \code{GTK_PRINT_SETTINGS_RESOLUTION},
\code{GTK_PRINT_SETTINGS_RESOLUTION_X} and 
\code{GTK_PRINT_SETTINGS_RESOLUTION_Y}.}
\usage{gtkPrintSettingsSetResolution(object, resolution)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{resolution}}{the resolution in dpi}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
