\alias{gtkPrintSettingsSetReverse}
\name{gtkPrintSettingsSetReverse}
\title{gtkPrintSettingsSetReverse}
\description{Sets the value of \code{GTK_PRINT_SETTINGS_REVERSE}.}
\usage{gtkPrintSettingsSetReverse(object, reverse)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{reverse}}{whether to reverse the output}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
