\alias{gtkPrintSettingsUnset}
\name{gtkPrintSettingsUnset}
\title{gtkPrintSettingsUnset}
\description{Removes any value associated with \code{key}. 
This has the same effect as setting the value to \code{NULL}.}
\usage{gtkPrintSettingsUnset(object, key)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkPrintSettings}}}
\item{\verb{key}}{a key}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
