\alias{gtkProgressBarGetOrientation}
\name{gtkProgressBarGetOrientation}
\title{gtkProgressBarGetOrientation}
\description{Retrieves the current progress bar orientation.}
\usage{gtkProgressBarGetOrientation(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkProgressBar}}}}
\value{[\code{\link{GtkProgressBarOrientation}}] orientation of the progress bar}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
