\alias{gtkProgressBarNew}
\name{gtkProgressBarNew}
\title{gtkProgressBarNew}
\description{Creates a new \verb{\link{GtkProgressBar}}.}
\usage{gtkProgressBarNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}] a \verb{\link{GtkProgressBar}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
