\alias{gtkProgressBarNew}
\name{gtkProgressBarNew}
\title{gtkProgressBarNew}
\description{Creates a new \code{\link{GtkProgressBar}}.}
\usage{gtkProgressBarNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] a \code{\link{GtkProgressBar}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
