\alias{gtkProgressBarNewWithAdjustment}
\name{gtkProgressBarNewWithAdjustment}
\title{gtkProgressBarNewWithAdjustment}
\description{
Creates a new \code{\link{GtkProgressBar}} with an associated \code{\link{GtkAdjustment}}.
\strong{WARNING: \code{gtk_progress_bar_new_with_adjustment} is deprecated and should not be used in newly-written code.  
Creates a new \code{\link{GtkProgressBar}} with an associated \code{\link{GtkAdjustment}}.}
}
\usage{gtkProgressBarNewWithAdjustment(adjustment = NULL, show = TRUE)}
\arguments{\item{\code{adjustment}}{[\code{\link{GtkAdjustment}}] a \code{\link{GtkAdjustment}}.}}
\value{[\code{\link{GtkWidget}}] a \code{\link{GtkProgressBar}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
