\alias{gtkProgressBarSetActivityBlocks}
\name{gtkProgressBarSetActivityBlocks}
\title{gtkProgressBarSetActivityBlocks}
\description{
Sets the number of blocks used when the progress bar is in activity
mode.  Larger numbers make the visible block smaller.
\strong{WARNING: \code{gtk_progress_bar_set_activity_blocks} is deprecated and should not be used in newly-written code.}
}
\usage{gtkProgressBarSetActivityBlocks(object, blocks)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkProgressBar}}.}
\item{\verb{blocks}}{number of blocks which can fit within the progress bar area.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
