\alias{gtkProgressBarSetBarStyle}
\name{gtkProgressBarSetBarStyle}
\title{gtkProgressBarSetBarStyle}
\description{
Sets the style of the \code{\link{GtkProgressBar}}.  The default style is
\code{GTK_PROGRESS_CONTINUOUS}.
\strong{WARNING: \code{gtk_progress_bar_set_bar_style} is deprecated and should not be used in newly-written code.}
}
\usage{gtkProgressBarSetBarStyle(object, style)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkProgressBar}}.}
\item{\verb{style}}{a \code{\link{GtkProgressBarStyle}} value indicating the desired style.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
