\alias{gtkProgressBarSetEllipsize}
\name{gtkProgressBarSetEllipsize}
\title{gtkProgressBarSetEllipsize}
\description{Sets the mode used to ellipsize (add an ellipsis: "...") the text 
if there is not enough space to render the entire string.}
\usage{gtkProgressBarSetEllipsize(object, mode)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkProgressBar}}}
\item{\verb{mode}}{a \code{\link{PangoEllipsizeMode}}}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
