\alias{gtkProgressBarSetOrientation}
\name{gtkProgressBarSetOrientation}
\title{gtkProgressBarSetOrientation}
\description{Causes the progress bar to switch to a different orientation
(left-to-right, right-to-left, top-to-bottom, or bottom-to-top).}
\usage{gtkProgressBarSetOrientation(object, orientation)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkProgressBar}}}
\item{\verb{orientation}}{orientation of the progress bar}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
