\alias{gtkProgressBarSetText}
\name{gtkProgressBarSetText}
\title{gtkProgressBarSetText}
\description{Causes the given \code{text} to appear superimposed on the progress bar.}
\usage{gtkProgressBarSetText(object, text)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkProgressBar}}]  a \verb{\link{GtkProgressBar}}}
\item{\verb{text}}{[character]  a UTF-8 string, or \code{NULL} }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
