\alias{gtkProgressBarSetText}
\name{gtkProgressBarSetText}
\title{gtkProgressBarSetText}
\description{Causes the given \code{text} to appear superimposed on the progress bar.}
\usage{gtkProgressBarSetText(object, text)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkProgressBar}}}
\item{\verb{text}}{a UTF-8 string, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
