\alias{gtkProgressConfigure}
\name{gtkProgressConfigure}
\title{gtkProgressConfigure}
\description{
Allows the configuration of the minimum, maximum, and current values for
the \code{\link{GtkProgress}}.
\strong{WARNING: \code{gtk_progress_configure} is deprecated and should not be used in newly-written code.}
}
\usage{gtkProgressConfigure(object, value, min, max)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkProgress}}.}
\item{\verb{value}}{the current progress value.}
\item{\verb{min}}{the minimum progress value.}
\item{\verb{max}}{the maximum progress value.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
