\alias{gtkProgressSetPercentage}
\name{gtkProgressSetPercentage}
\title{gtkProgressSetPercentage}
\description{
Sets the current percentage completion for the \code{\link{GtkProgress}}.
\strong{WARNING: \code{gtk_progress_set_percentage} is deprecated and should not be used in newly-written code.}
}
\usage{gtkProgressSetPercentage(object, percentage)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkProgress}}.}
\item{\verb{percentage}}{the percentage complete which must be between 0.0
and 1.0.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
