\alias{gtkProgressSetTextAlignment}
\name{gtkProgressSetTextAlignment}
\title{gtkProgressSetTextAlignment}
\description{
Controls the alignment of the text within the progress bar area.
\strong{WARNING: \code{gtk_progress_set_text_alignment} is deprecated and should not be used in newly-written code.}
}
\usage{gtkProgressSetTextAlignment(object, x.align, y.align)}
\arguments{
\item{\code{object}}{[\code{\link{GtkProgress}}] a \code{\link{GtkProgress}}.}
\item{\code{x.align}}{[numeric] a number between 0.0 and 1.0 indicating the horizontal
alignment of the progress text within the \code{\link{GtkProgress}}.}
\item{\code{y.align}}{[numeric] a number between 0.0 and 1.0 indicating the vertical
alignment of the progress text within the \code{\link{GtkProgress}}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
