\alias{gtkQuitAddDestroy}
\name{gtkQuitAddDestroy}
\title{gtkQuitAddDestroy}
\description{Trigger destruction of \code{object} in case the mainloop at level \code{main.level}
is quit.}
\usage{gtkQuitAddDestroy(main.level, object)}
\arguments{
\item{\verb{main.level}}{Level of the mainloop which shall trigger the destruction.}
\item{\verb{object}}{Object to be destroyed.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
