\alias{gtkQuitRemove}
\name{gtkQuitRemove}
\title{gtkQuitRemove}
\description{Removes a quit handler by its identifier.}
\usage{gtkQuitRemove(quit.handler.id)}
\arguments{\item{\verb{quit.handler.id}}{Identifier for the handler returned when installing it.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
