\alias{gtkQuitRemoveByData}
\name{gtkQuitRemoveByData}
\title{gtkQuitRemoveByData}
\description{Removes a quit handler identified by its \code{data} field.}
\usage{gtkQuitRemoveByData(data)}
\arguments{\item{\verb{data}}{The pointer passed as \code{data} to \code{\link{gtkQuitAdd}} or \code{\link{gtkQuitAddFull}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
