\alias{gtkRadioActionSetCurrentValue}
\name{gtkRadioActionSetCurrentValue}
\title{gtkRadioActionSetCurrentValue}
\description{Sets the currently active group member to the member with value
property \code{current.value}.}
\usage{gtkRadioActionSetCurrentValue(object, current.value)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkRadioAction}}]  a \verb{\link{GtkRadioAction}}}
\item{\verb{current.value}}{[integer]  the new value}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
