\alias{gtkRadioActionSetCurrentValue}
\name{gtkRadioActionSetCurrentValue}
\title{gtkRadioActionSetCurrentValue}
\description{Sets the currently active group member to the member with value
property \code{current.value}.}
\usage{gtkRadioActionSetCurrentValue(object, current.value)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRadioAction}}}
\item{\verb{current.value}}{the new value}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
