\alias{gtkRadioActionSetGroup}
\name{gtkRadioActionSetGroup}
\title{gtkRadioActionSetGroup}
\description{Sets the radio group for the radio action object.}
\usage{gtkRadioActionSetGroup(object, group)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkRadioAction}}]  the action object}
\item{\verb{group}}{[list]  a list representing a radio group}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
