\alias{gtkRadioActionSetGroup}
\name{gtkRadioActionSetGroup}
\title{gtkRadioActionSetGroup}
\description{Sets the radio group for the radio action object.}
\usage{gtkRadioActionSetGroup(object, group)}
\arguments{
\item{\verb{object}}{the action object}
\item{\verb{group}}{a list representing a radio group}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
