\alias{gtkRadioButtonGetGroup}
\name{gtkRadioButtonGetGroup}
\title{gtkRadioButtonGetGroup}
\description{Retrieves the group assigned to a radio button.}
\usage{gtkRadioButtonGetGroup(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkRadioButton}}] a \verb{\link{GtkRadioButton}}.}}
\value{[list] a linked list containing all the radio buttons in the same group
as \code{radio.button}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
