\alias{gtkRadioButtonGetGroup}
\name{gtkRadioButtonGetGroup}
\title{gtkRadioButtonGetGroup}
\description{Retrieves the group assigned to a radio button.}
\usage{gtkRadioButtonGetGroup(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRadioButton}}.}}
\value{[list] a linked list
containing all the radio buttons in the same group
as \code{radio.button}. \emph{[  \acronym{element-type}   GtkRadioButton][  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
