\alias{gtkRadioButtonNewFromWidget}
\name{gtkRadioButtonNewFromWidget}
\title{gtkRadioButtonNewFromWidget}
\description{Creates a new \code{\link{GtkRadioButton}}, adding it to the same group as \code{radio.group.member}.
As with \code{\link{gtkRadioButtonNew}}, a widget should be packed into the radio button.}
\usage{gtkRadioButtonNewFromWidget(group = NULL, show = TRUE)}
\arguments{\item{\verb{group}}{an existing \code{\link{GtkRadioButton}}.}}
\value{[\code{\link{GtkWidget}}] a new radio button.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
