\alias{gtkRadioButtonNewWithLabelFromWidget}
\name{gtkRadioButtonNewWithLabelFromWidget}
\title{gtkRadioButtonNewWithLabelFromWidget}
\description{Creates a new \code{\link{GtkRadioButton}} with a text label, adding it to the same group
as \code{group}.}
\usage{gtkRadioButtonNewWithLabelFromWidget(group = NULL, label, show = TRUE)}
\arguments{
\item{\code{group}}{[\code{\link{GtkRadioButton}}] }
\item{\code{label}}{[character] a text string to display next to the radio button.}
}
\value{[\code{\link{GtkWidget}}] a new radio button.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
