\alias{gtkRadioButtonNewWithLabelFromWidget}
\name{gtkRadioButtonNewWithLabelFromWidget}
\title{gtkRadioButtonNewWithLabelFromWidget}
\description{Creates a new \code{\link{GtkRadioButton}} with a text label, adding it to the same group
as \code{radio.group.member}.}
\usage{gtkRadioButtonNewWithLabelFromWidget(group = NULL, label, show = TRUE)}
\arguments{
\item{\verb{group}}{widget to get radio group from or NULL}
\item{\verb{label}}{a text string to display next to the radio button.}
}
\value{[\code{\link{GtkWidget}}] a new radio button.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
