\alias{gtkRadioMenuItemGetGroup}
\name{gtkRadioMenuItemGetGroup}
\title{gtkRadioMenuItemGetGroup}
\description{Returns the group to which the radio menu item belongs, as a \code{list} of
\code{\link{GtkRadioMenuItem}}.}
\usage{gtkRadioMenuItemGetGroup(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkRadioMenuItem}}] a \code{\link{GtkRadioMenuItem}}.}}
\value{[list] the group of \code{radio.menu.item}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
