\alias{gtkRadioMenuItemGetGroup}
\name{gtkRadioMenuItemGetGroup}
\title{gtkRadioMenuItemGetGroup}
\description{Returns the group to which the radio menu item belongs, as a \verb{list} of
\code{\link{GtkRadioMenuItem}}. The list belongs to GTK+ and should not be freed.}
\usage{gtkRadioMenuItemGetGroup(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRadioMenuItem}}.}}
\value{[list] the group of \code{radio.menu.item}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
