\alias{gtkRadioMenuItemNew}
\name{gtkRadioMenuItemNew}
\title{gtkRadioMenuItemNew}
\description{Creates a new \code{\link{GtkRadioMenuItem}}.}
\usage{gtkRadioMenuItemNew(group = NULL, show = TRUE)}
\arguments{\item{\verb{group}}{the group to which the radio menu item is to be attached}}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkRadioMenuItem}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
