\alias{gtkRadioMenuItemNewFromWidget}
\name{gtkRadioMenuItemNewFromWidget}
\title{gtkRadioMenuItemNewFromWidget}
\description{Creates a new \code{\link{GtkRadioMenuItem}} adding it to the same group as \code{group}.}
\usage{gtkRadioMenuItemNewFromWidget(group = NULL, show = TRUE)}
\arguments{\item{\verb{group}}{An existing \code{\link{GtkRadioMenuItem}}}}
\details{Since 2.4}
\value{[\code{\link{GtkWidget}}] The new \code{\link{GtkRadioMenuItem}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
