\alias{gtkRadioMenuItemNewWithLabel}
\name{gtkRadioMenuItemNewWithLabel}
\title{gtkRadioMenuItemNewWithLabel}
\description{Creates a new \verb{\link{GtkRadioMenuItem}} whose child is a simple \verb{\link{GtkLabel}}.}
\usage{gtkRadioMenuItemNewWithLabel(group = NULL, label, show = TRUE)}
\arguments{
\item{\verb{group}}{[list] the group to which the radio menu item is to be attached}
\item{\verb{label}}{[character] the text for the label}
}
\value{[\verb{\link{GtkWidget}}] a new \verb{\link{GtkRadioMenuItem}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
