\alias{gtkRadioMenuItemNewWithLabel}
\name{gtkRadioMenuItemNewWithLabel}
\title{gtkRadioMenuItemNewWithLabel}
\description{Creates a new \code{\link{GtkRadioMenuItem}} whose child is a simple \code{\link{GtkLabel}}.}
\usage{gtkRadioMenuItemNewWithLabel(group = NULL, label, show = TRUE)}
\arguments{
\item{\verb{group}}{. \emph{[  \acronym{element-type}   GtkRadioMenuItem][  \acronym{transfer full}  ]}}
\item{\verb{label}}{the text for the label}
}
\value{[\code{\link{GtkWidget}}] A new \code{\link{GtkRadioMenuItem}}. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
