\alias{gtkRadioMenuItemNewWithLabelFromWidget}
\name{gtkRadioMenuItemNewWithLabelFromWidget}
\title{gtkRadioMenuItemNewWithLabelFromWidget}
\description{Creates a new GtkRadioMenuItem whose child is a simple GtkLabel.
The new \code{\link{GtkRadioMenuItem}} is added to the same group as \code{group}.}
\usage{gtkRadioMenuItemNewWithLabelFromWidget(group = NULL, label, show = TRUE)}
\arguments{
\item{\verb{group}}{an existing \code{\link{GtkRadioMenuItem}}}
\item{\verb{label}}{the text for the label}
}
\details{Since 2.4}
\value{[\code{\link{GtkWidget}}] The new \code{\link{GtkRadioMenuItem}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
