\alias{gtkRadioMenuItemSetGroup}
\name{gtkRadioMenuItemSetGroup}
\title{gtkRadioMenuItemSetGroup}
\description{Sets the group of a radio menu item, or changes it.}
\usage{gtkRadioMenuItemSetGroup(object, group)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRadioMenuItem}}.}
\item{\verb{group}}{the new group.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
