\alias{gtkRadioToolButtonGetGroup}
\name{gtkRadioToolButtonGetGroup}
\title{gtkRadioToolButtonGetGroup}
\description{Returns the radio button group \code{button} belongs to.}
\usage{gtkRadioToolButtonGetGroup(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRadioToolButton}}}}
\details{Since 2.4}
\value{[list] The group \code{button} belongs to.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
