\alias{gtkRadioToolButtonNew}
\name{gtkRadioToolButtonNew}
\title{gtkRadioToolButtonNew}
\description{Creates a new \code{\link{GtkRadioToolButton}}, adding it to \code{group}.}
\usage{gtkRadioToolButtonNew(group = NULL, show = TRUE)}
\arguments{\item{\verb{group}}{An existing radio button group, or \code{NULL} if you are creating a new group. \emph{[  \acronym{allow-none}  ]}}}
\details{Since 2.4}
\value{[\code{\link{GtkToolItem}}] The new \code{\link{GtkRadioToolButton}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
