\alias{gtkRadioToolButtonNewFromWidget}
\name{gtkRadioToolButtonNewFromWidget}
\title{gtkRadioToolButtonNewFromWidget}
\description{Creates a new \code{\link{GtkRadioToolButton}} adding it to the same group as \code{gruup}}
\usage{gtkRadioToolButtonNewFromWidget(group = NULL, show = TRUE)}
\arguments{\item{\verb{group}}{An existing \code{\link{GtkRadioToolButton}}}}
\details{Since 2.4}
\value{[\code{\link{GtkToolItem}}] The new \code{\link{GtkRadioToolButton}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
