\alias{gtkRangeGetAdjustment}
\name{gtkRangeGetAdjustment}
\title{gtkRangeGetAdjustment}
\description{Get the \code{\link{GtkAdjustment}} which is the "model" object for \code{\link{GtkRange}}.
See \code{\link{gtkRangeSetAdjustment}} for details.
The return value does not have a reference added, so should not
be unreferenced.}
\usage{gtkRangeGetAdjustment(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRange}}}}
\value{[\code{\link{GtkAdjustment}}] a \code{\link{GtkAdjustment}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
