\alias{gtkRangeGetFillLevel}
\name{gtkRangeGetFillLevel}
\title{gtkRangeGetFillLevel}
\description{Gets the current position of the fill level indicator.}
\usage{gtkRangeGetFillLevel(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkRange}}}}
\details{Since 2.12}
\value{[numeric] The current fill level}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
