\alias{gtkRangeGetInverted}
\name{gtkRangeGetInverted}
\title{gtkRangeGetInverted}
\description{Gets the value set by \code{\link{gtkRangeSetInverted}}.}
\usage{gtkRangeGetInverted(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRange}}}}
\value{[logical] \code{TRUE} if the range is inverted}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
