\alias{gtkRangeGetLowerStepperSensitivity}
\name{gtkRangeGetLowerStepperSensitivity}
\title{gtkRangeGetLowerStepperSensitivity}
\description{Gets the sensitivity policy for the stepper that points to the
'lower' end of the GtkRange's adjustment.}
\usage{gtkRangeGetLowerStepperSensitivity(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRange}}}}
\details{Since 2.10}
\value{[\code{\link{GtkSensitivityType}}] The lower stepper's sensitivity policy.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
