\alias{gtkRangeGetShowFillLevel}
\name{gtkRangeGetShowFillLevel}
\title{gtkRangeGetShowFillLevel}
\description{Gets whether the range displays the fill level graphically.}
\usage{gtkRangeGetShowFillLevel(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkRange}}}}
\details{Since 2.12}
\value{[logical] \code{TRUE} if \code{range} shows the fill level.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
