\alias{gtkRangeGetUpdatePolicy}
\name{gtkRangeGetUpdatePolicy}
\title{gtkRangeGetUpdatePolicy}
\description{Gets the update policy of \code{range}. See \code{\link{gtkRangeSetUpdatePolicy}}.}
\usage{gtkRangeGetUpdatePolicy(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRange}}}}
\value{[\code{\link{GtkUpdateType}}] the current update policy}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
