\alias{gtkRangeSetInverted}
\name{gtkRangeSetInverted}
\title{gtkRangeSetInverted}
\description{Ranges normally move from lower to higher values as the
slider moves from top to bottom or left to right. Inverted
ranges have higher values at the top or on the right rather than
on the bottom or left.}
\usage{gtkRangeSetInverted(object, setting)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRange}}}
\item{\verb{setting}}{\code{TRUE} to invert the range}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
