\alias{gtkRangeSetLowerStepperSensitivity}
\name{gtkRangeSetLowerStepperSensitivity}
\title{gtkRangeSetLowerStepperSensitivity}
\description{Sets the sensitivity policy for the stepper that points to the
'lower' end of the GtkRange's adjustment.}
\usage{gtkRangeSetLowerStepperSensitivity(object, sensitivity)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRange}}}
\item{\verb{sensitivity}}{the lower stepper's sensitivity policy.}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
