\alias{gtkRangeSetUpperStepperSensitivity}
\name{gtkRangeSetUpperStepperSensitivity}
\title{gtkRangeSetUpperStepperSensitivity}
\description{Sets the sensitivity policy for the stepper that points to the
'upper' end of the GtkRange's adjustment.}
\usage{gtkRangeSetUpperStepperSensitivity(object, sensitivity)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRange}}}
\item{\verb{sensitivity}}{the upper stepper's sensitivity policy.}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
