\alias{gtkRcGetModuleDir}
\name{gtkRcGetModuleDir}
\title{gtkRcGetModuleDir}
\description{Returns a directory in which GTK+ looks for theme engines.
For full information about the search for theme engines,
see the docs for \env{GTK_PATH} in
.}
\usage{gtkRcGetModuleDir()}
\value{[character] the directory.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
