\alias{gtkRcGetThemeDir}
\name{gtkRcGetThemeDir}
\title{gtkRcGetThemeDir}
\description{Returns the standard directory in which themes should
be installed. (GTK+ does not actually use this directory
itself.)}
\usage{gtkRcGetThemeDir()}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
