\alias{gtkRcStyleCopy}
\name{gtkRcStyleCopy}
\title{gtkRcStyleCopy}
\description{Makes a copy of the specified \code{\link{GtkRcStyle}}. This function
will correctly copy an RC style that is a member of a class
derived from \code{\link{GtkRcStyle}}.}
\usage{gtkRcStyleCopy(object)}
\arguments{\item{\verb{object}}{the style to copy}}
\value{[\code{\link{GtkRcStyle}}] the resulting \code{\link{GtkRcStyle}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
