\alias{gtkRcStyleNew}
\name{gtkRcStyleNew}
\title{gtkRcStyleNew}
\description{Creates a new \verb{\link{GtkRcStyle}} with no fields set and
a reference count of 1.}
\usage{gtkRcStyleNew()}
\value{[\verb{\link{GtkRcStyle}}] the newly-created \verb{\link{GtkRcStyle}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
