\alias{gtkRecentActionGetShowNumbers}
\name{gtkRecentActionGetShowNumbers}
\title{gtkRecentActionGetShowNumbers}
\description{Returns the value set by \code{\link{gtkRecentChooserMenuSetShowNumbers}}.}
\usage{gtkRecentActionGetShowNumbers(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentAction}}}}
\details{Since 2.12}
\value{[logical] \code{TRUE} if numbers should be shown.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
