\alias{gtkRecentChooserGetCurrentItem}
\name{gtkRecentChooserGetCurrentItem}
\title{gtkRecentChooserGetCurrentItem}
\description{Gets the \code{\link{GtkRecentInfo}} currently selected by \code{chooser}.}
\usage{gtkRecentChooserGetCurrentItem(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}}
\details{Since 2.10}
\value{[\code{\link{GtkRecentInfo}}] a \code{\link{GtkRecentInfo}}.  Use \code{\link{gtkRecentInfoUnref}} when
when you have finished using it.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
