\alias{gtkRecentChooserGetFilter}
\name{gtkRecentChooserGetFilter}
\title{gtkRecentChooserGetFilter}
\description{Gets the \code{\link{GtkRecentFilter}} object currently used by \code{chooser} to affect
the display of the recently used resources.}
\usage{gtkRecentChooserGetFilter(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}}
\details{Since 2.10}
\value{[\code{\link{GtkRecentFilter}}] a \code{\link{GtkRecentFilter}} object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
