\alias{gtkRecentChooserGetSelectMultiple}
\name{gtkRecentChooserGetSelectMultiple}
\title{gtkRecentChooserGetSelectMultiple}
\description{Gets whether \code{chooser} can select multiple items.}
\usage{gtkRecentChooserGetSelectMultiple(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}}
\details{Since 2.10}
\value{[logical] \code{TRUE} if \code{chooser} can select more than one item.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
