\alias{gtkRecentChooserGetSortType}
\name{gtkRecentChooserGetSortType}
\title{gtkRecentChooserGetSortType}
\description{Gets the value set by \code{\link{gtkRecentChooserSetSortType}}.}
\usage{gtkRecentChooserGetSortType(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}}
\details{Since 2.10}
\value{[\code{\link{GtkRecentSortType}}] the sorting order of the \code{chooser}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
