\alias{gtkRecentChooserListFilters}
\name{gtkRecentChooserListFilters}
\title{gtkRecentChooserListFilters}
\description{Gets the \code{\link{GtkRecentFilter}} objects held by \code{chooser}.}
\usage{gtkRecentChooserListFilters(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}}
\details{Since 2.10}
\value{[list] A singly linked list
of \code{\link{GtkRecentFilter}} objects. \emph{[  \acronym{element-type}   GtkRecentFilter][  \acronym{transfer container}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
