\alias{gtkRecentChooserRemoveFilter}
\name{gtkRecentChooserRemoveFilter}
\title{gtkRecentChooserRemoveFilter}
\description{Removes \code{filter} from the list of \verb{\link{GtkRecentFilter}} objects held by \code{chooser}.}
\usage{gtkRecentChooserRemoveFilter(object, filter)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkRecentChooser}}]  a \verb{\link{GtkRecentChooser}}}
\item{\verb{filter}}{[\verb{\link{GtkRecentFilter}}]  a \verb{\link{GtkRecentFilter}}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
