\alias{gtkRecentChooserRemoveFilter}
\name{gtkRecentChooserRemoveFilter}
\title{gtkRecentChooserRemoveFilter}
\description{Removes \code{filter} from the list of \code{\link{GtkRecentFilter}} objects held by \code{chooser}.}
\usage{gtkRecentChooserRemoveFilter(object, filter)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}
\item{\verb{filter}}{a \code{\link{GtkRecentFilter}}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
