\alias{gtkRecentChooserSetLimit}
\name{gtkRecentChooserSetLimit}
\title{gtkRecentChooserSetLimit}
\description{Sets the number of items that should be returned by
\code{\link{gtkRecentChooserGetItems}} and \code{\link{gtkRecentChooserGetUris}}.}
\usage{gtkRecentChooserSetLimit(object, limit)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}
\item{\verb{limit}}{a positive integer, or -1 for all items}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
