\alias{gtkRecentChooserSetSelectMultiple}
\name{gtkRecentChooserSetSelectMultiple}
\title{gtkRecentChooserSetSelectMultiple}
\description{Sets whether \code{chooser} can select multiple items.}
\usage{gtkRecentChooserSetSelectMultiple(object, select.multiple)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}
\item{\verb{select.multiple}}{\code{TRUE} if \code{chooser} can select more than one item}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
