\alias{gtkRecentChooserSetShowIcons}
\name{gtkRecentChooserSetShowIcons}
\title{gtkRecentChooserSetShowIcons}
\description{Sets whether \code{chooser} should show an icon near the resource when
displaying it.}
\usage{gtkRecentChooserSetShowIcons(object, show.icons)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}
\item{\verb{show.icons}}{whether to show an icon near the resource}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
