\alias{gtkRecentChooserSetShowTips}
\name{gtkRecentChooserSetShowTips}
\title{gtkRecentChooserSetShowTips}
\description{Sets whether to show a tooltips on the widget.}
\usage{gtkRecentChooserSetShowTips(object, show.tips)}
\arguments{
\item{\code{object}}{[\code{\link{GtkRecentChooser}}]  a \code{\link{GtkRecentChooser}}}
\item{\code{show.tips}}{[logical]  \code{TRUE} if tooltips should be shown}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
