\alias{gtkRecentChooserSetShowTips}
\name{gtkRecentChooserSetShowTips}
\title{gtkRecentChooserSetShowTips}
\description{Sets whether to show a tooltips containing the full path of each
recently used resource in a \code{\link{GtkRecentChooser}} widget.}
\usage{gtkRecentChooserSetShowTips(object, show.tips)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}
\item{\verb{show.tips}}{\code{TRUE} if tooltips should be shown}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
