\alias{gtkRecentChooserSetSortType}
\name{gtkRecentChooserSetSortType}
\title{gtkRecentChooserSetSortType}
\description{Changes the sorting order of the recently used resources list displayed by
\code{chooser}.}
\usage{gtkRecentChooserSetSortType(object, sort.type)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}
\item{\verb{sort.type}}{sort order that the chooser should use}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
