\alias{gtkRecentChooserUnselectAll}
\name{gtkRecentChooserUnselectAll}
\title{gtkRecentChooserUnselectAll}
\description{Unselects all the items inside \code{chooser}.}
\usage{gtkRecentChooserUnselectAll(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
