\alias{gtkRecentChooserUnselectUri}
\name{gtkRecentChooserUnselectUri}
\title{gtkRecentChooserUnselectUri}
\description{Unselects \code{uri} inside \code{chooser}.}
\usage{gtkRecentChooserUnselectUri(object, uri)}
\arguments{
\item{\code{object}}{[\code{\link{GtkRecentChooser}}]  a \code{\link{GtkRecentChooser}}}
\item{\code{uri}}{[character]  a URI}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
