\alias{gtkRecentChooserUnselectUri}
\name{gtkRecentChooserUnselectUri}
\title{gtkRecentChooserUnselectUri}
\description{Unselects \code{uri} inside \code{chooser}.}
\usage{gtkRecentChooserUnselectUri(object, uri)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentChooser}}}
\item{\verb{uri}}{a URI}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
