\alias{gtkRecentChooserWidgetNew}
\name{gtkRecentChooserWidgetNew}
\title{gtkRecentChooserWidgetNew}
\description{Creates a new \code{\link{GtkRecentChooserWidget}} object.  This is an embeddable widget
used to access the recently used resources list.}
\usage{gtkRecentChooserWidgetNew()}
\details{Since 2.10}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkRecentChooserWidget}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
