\alias{gtkRecentFilterAddAge}
\name{gtkRecentFilterAddAge}
\title{gtkRecentFilterAddAge}
\description{Adds a rule that allows resources based on their age - that is, the number
of days elapsed since they were last modified.}
\usage{gtkRecentFilterAddAge(object, days)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentFilter}}}
\item{\verb{days}}{number of days}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
