\alias{gtkRecentFilterAddApplication}
\name{gtkRecentFilterAddApplication}
\title{gtkRecentFilterAddApplication}
\description{Adds a rule that allows resources based on the name of the application
that has registered them.}
\usage{gtkRecentFilterAddApplication(object, application)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentFilter}}}
\item{\verb{application}}{an application name}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
