\alias{gtkRecentFilterAddGroup}
\name{gtkRecentFilterAddGroup}
\title{gtkRecentFilterAddGroup}
\description{Adds a rule that allows resources based on the name of the group
to which they belong}
\usage{gtkRecentFilterAddGroup(object, group)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentFilter}}}
\item{\verb{group}}{a group name}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
