\alias{gtkRecentFilterAddMimeType}
\name{gtkRecentFilterAddMimeType}
\title{gtkRecentFilterAddMimeType}
\description{Adds a rule that allows resources based on their registered MIME type.}
\usage{gtkRecentFilterAddMimeType(object, mime.type)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentFilter}}}
\item{\verb{mime.type}}{a MIME type}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
