\alias{gtkRecentFilterAddPattern}
\name{gtkRecentFilterAddPattern}
\title{gtkRecentFilterAddPattern}
\description{Adds a rule that allows resources based on a pattern matching their
display name.}
\usage{gtkRecentFilterAddPattern(object, pattern)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentFilter}}}
\item{\verb{pattern}}{a file pattern}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
