\alias{gtkRecentFilterGetName}
\name{gtkRecentFilterGetName}
\title{gtkRecentFilterGetName}
\description{Gets the human-readable name for the filter.
See \code{\link{gtkRecentFilterSetName}}.}
\usage{gtkRecentFilterGetName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentFilter}}}}
\details{Since 2.10}
\value{[character] the name of the filter, or \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
