\alias{gtkRecentInfoGetAge}
\name{gtkRecentInfoGetAge}
\title{gtkRecentInfoGetAge}
\description{Gets the number of days elapsed since the last update of the resource
pointed by \code{info}.}
\usage{gtkRecentInfoGetAge(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentInfo}}}}
\details{Since 2.10}
\value{[integer] a positive integer containing the number of days elapsed
since the time this resource was last modified.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
