\alias{gtkRecentInfoGetApplications}
\name{gtkRecentInfoGetApplications}
\title{gtkRecentInfoGetApplications}
\description{Retrieves the list of applications that have registered this resource.}
\usage{gtkRecentInfoGetApplications(object, length)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkRecentInfo}}]  a \verb{\link{GtkRecentInfo}}}
\item{\verb{length}}{[numeric]  return location for the length of the returned list, or \code{NULL}}
}
\details{  Since  2.10}
\value{[character]  a newly allocated list of strings.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
