\alias{gtkRecentInfoGetApplications}
\name{gtkRecentInfoGetApplications}
\title{gtkRecentInfoGetApplications}
\description{Retrieves the list of applications that have registered this resource.}
\usage{gtkRecentInfoGetApplications(object, length)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentInfo}}}
\item{\verb{length}}{return location for the length of the returned list. \emph{[  \acronym{out}  ][  \acronym{allow-none}  ]}}
}
\details{Since 2.10}
\value{[character] a newly list of strings. \emph{[  \acronym{array}   length=length zero-terminated=1]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
