\alias{gtkRecentInfoGetDescription}
\name{gtkRecentInfoGetDescription}
\title{gtkRecentInfoGetDescription}
\description{Gets the (short) description of the resource.}
\usage{gtkRecentInfoGetDescription(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentInfo}}}}
\details{Since 2.10}
\value{[character] the description of the resource. and should not be freed.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
