\alias{gtkRecentInfoGetMimeType}
\name{gtkRecentInfoGetMimeType}
\title{gtkRecentInfoGetMimeType}
\description{Gets the MIME type of the resource.}
\usage{gtkRecentInfoGetMimeType(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentInfo}}}}
\details{Since 2.10}
\value{[character] the MIME type of the resource. and should not be freed.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
