\alias{gtkRecentInfoGetShortName}
\name{gtkRecentInfoGetShortName}
\title{gtkRecentInfoGetShortName}
\description{Computes a valid UTF-8 string that can be used as the name of the item in a
menu or list.  For example, calling this function on an item that refers to
"file:///foo/bar.txt" will yield "bar.txt".}
\usage{gtkRecentInfoGetShortName(object)}
\arguments{\item{\verb{object}}{an \code{\link{GtkRecentInfo}}}}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
