\alias{gtkRecentInfoGetUri}
\name{gtkRecentInfoGetUri}
\title{gtkRecentInfoGetUri}
\description{Gets the URI of the resource.}
\usage{gtkRecentInfoGetUri(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkRecentInfo}}]  a \verb{\link{GtkRecentInfo}}}}
\details{  Since  2.10}
\value{[character]  the URI of the resource.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
