\alias{gtkRecentInfoMatch}
\name{gtkRecentInfoMatch}
\title{gtkRecentInfoMatch}
\description{Checks whether two \code{\link{GtkRecentInfo}} structures point to the same
resource.}
\usage{gtkRecentInfoMatch(object, info.b)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentInfo}}}
\item{\verb{info.b}}{a \code{\link{GtkRecentInfo}}}
}
\details{Since 2.10}
\value{[logical] \code{TRUE} if both \code{\link{GtkRecentInfo}} structures point to se same
resource, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
