\alias{gtkRecentInfoRef}
\name{gtkRecentInfoRef}
\title{gtkRecentInfoRef}
\description{Increases the reference count of \code{recent.info} by one.}
\usage{gtkRecentInfoRef(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentInfo}}}}
\details{Since 2.10}
\value{[\code{\link{GtkRecentInfo}}] the recent info object with its reference count increased
by one.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
