\alias{gtkRecentInfoUnref}
\name{gtkRecentInfoUnref}
\title{gtkRecentInfoUnref}
\description{Decreases the reference count of \code{info} by one.  If the reference
count reaches zero, \code{info} is deallocated, and the memory freed.}
\usage{gtkRecentInfoUnref(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRecentInfo}}}}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
