\alias{gtkRecentManagerGetDefault}
\name{gtkRecentManagerGetDefault}
\title{gtkRecentManagerGetDefault}
\description{Gets a unique instance of \code{\link{GtkRecentManager}}, that you can share
in your application without caring about memory management.}
\usage{gtkRecentManagerGetDefault()}
\details{Since 2.10}
\value{[\code{\link{GtkRecentManager}}] A unique \code{\link{GtkRecentManager}}. Do not ref or unref it. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
