\alias{gtkRecentManagerHasItem}
\name{gtkRecentManagerHasItem}
\title{gtkRecentManagerHasItem}
\description{Checks whether there is a recently used resource registered
with \code{uri} inside the recent manager.}
\usage{gtkRecentManagerHasItem(object, uri)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentManager}}}
\item{\verb{uri}}{a URI}
}
\details{Since 2.10}
\value{[logical] \code{TRUE} if the resource was found, \code{FALSE} otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
