\alias{gtkRecentManagerSetScreen}
\name{gtkRecentManagerSetScreen}
\title{gtkRecentManagerSetScreen}
\description{
Sets the screen for a recent manager; the screen is used to
track the user's currently configured recently used documents
storage.
\strong{WARNING: \code{gtk_recent_manager_set_screen} has been deprecated since version 2.12 and should not be used in newly-written code. This function has been deprecated and should
  not be used in newly written code. Calling this function has
  no effect.}
}
\usage{gtkRecentManagerSetScreen(object, screen)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkRecentManager}}}
\item{\verb{screen}}{a \code{\link{GdkScreen}}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
