\alias{gtkRequisitionCopy}
\name{gtkRequisitionCopy}
\title{gtkRequisitionCopy}
\description{Copies a \verb{\link{GtkRequisition}}.}
\usage{gtkRequisitionCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkRequisition}}]  a \verb{\link{GtkRequisition}}}}
\value{[\verb{\link{GtkRequisition}}]  a copy of \code{requisition}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
