\alias{gtkRequisitionCopy}
\name{gtkRequisitionCopy}
\title{gtkRequisitionCopy}
\description{Copies a \code{\link{GtkRequisition}}.}
\usage{gtkRequisitionCopy(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRequisition}}}}
\value{[\code{\link{GtkRequisition}}] a copy of \code{requisition}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
