\alias{gtkRulerGetMetric}
\name{gtkRulerGetMetric}
\title{gtkRulerGetMetric}
\description{Gets the units used for a \code{\link{GtkRuler}}. See \code{\link{gtkRulerSetMetric}}.}
\usage{gtkRulerGetMetric(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkRuler}}}}
\value{[\code{\link{GtkMetricType}}] the units currently used for \code{ruler}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
