\alias{gtkRulerSetMetric}
\name{gtkRulerSetMetric}
\title{gtkRulerSetMetric}
\description{This calls the \code{GTKMetricType} to set the ruler to units defined. Available units
are GTK\_PIXELS, GTK\_INCHES, or GTK\_CENTIMETERS. The default unit of measurement
is GTK\_PIXELS.}
\usage{gtkRulerSetMetric(object, metric)}
\arguments{
\item{\code{object}}{[\code{\link{GtkRuler}}] the gtkruler}
\item{\code{metric}}{[\code{\link{GtkMetricType}}] the unit of measurement}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
