\alias{gtkRulerSetMetric}
\name{gtkRulerSetMetric}
\title{gtkRulerSetMetric}
\description{This calls the \verb{GTKMetricType} to set the ruler to units defined. Available units
are GTK_PIXELS, GTK_INCHES, or GTK_CENTIMETERS. The default unit of measurement
is GTK_PIXELS.}
\usage{gtkRulerSetMetric(object, metric)}
\arguments{
\item{\verb{object}}{the gtkruler}
\item{\verb{metric}}{the unit of measurement}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
