\alias{gtkRulerSetRange}
\name{gtkRulerSetRange}
\title{gtkRulerSetRange}
\description{This sets the range of the ruler.}
\usage{gtkRulerSetRange(object, lower, upper, position, max.size)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkRuler}}]  the gtkruler}
\item{\verb{lower}}{[numeric]  the lower limit of the ruler}
\item{\verb{upper}}{[numeric]  the upper limit of the ruler}
\item{\verb{position}}{[numeric]  the mark on the ruler}
\item{\verb{max.size}}{[numeric]  the maximum size of the ruler used when calculating the space to
leave for the text}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
