\alias{gtkRulerSetRange}
\name{gtkRulerSetRange}
\title{gtkRulerSetRange}
\description{This sets the range of the ruler using gfloat lower, gfloat upper, gfloat position, and gfloat max\_size. }
\usage{gtkRulerSetRange(object, lower, upper, position, max.size)}
\arguments{
\item{\code{object}}{[\code{\link{GtkRuler}}] the gtkruler}
\item{\code{lower}}{[numeric] the upper limit of the ruler}
\item{\code{upper}}{[numeric] the lower limit of the ruler}
\item{\code{position}}{[numeric] the mark on the ruler}
\item{\code{max.size}}{[numeric] the maximum size of the ruler}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
