\alias{gtkScaleButtonGetAdjustment}
\name{gtkScaleButtonGetAdjustment}
\title{gtkScaleButtonGetAdjustment}
\description{Gets the \code{\link{GtkAdjustment}} associated with the \code{\link{GtkScaleButton}}'s scale.
See \code{\link{gtkRangeGetAdjustment}} for details.}
\usage{gtkScaleButtonGetAdjustment(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkScaleButton}}}}
\details{Since 2.12}
\value{[\code{\link{GtkAdjustment}}] the adjustment associated with the scale}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
