\alias{gtkScaleButtonSetAdjustment}
\name{gtkScaleButtonSetAdjustment}
\title{gtkScaleButtonSetAdjustment}
\description{Sets the \verb{\link{GtkAdjustment}} to be used as a model 
for the \verb{\link{GtkScaleButton}}'s scale.
See \code{\link{gtkRangeSetAdjustment}} for details.}
\usage{gtkScaleButtonSetAdjustment(object, adjustment)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkScaleButton}}]  a \verb{\link{GtkScaleButton}}}
\item{\verb{adjustment}}{[\verb{\link{GtkAdjustment}}]  a \verb{\link{GtkAdjustment}}}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
