\alias{gtkScaleGetLayout}
\name{gtkScaleGetLayout}
\title{gtkScaleGetLayout}
\description{Gets the \code{\link{PangoLayout}} used to display the scale.}
\usage{gtkScaleGetLayout(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkScale}}]  A \code{\link{GtkScale}}}}
\details{  Since  2.4}
\value{[\code{\link{PangoLayout}}]  the \code{\link{PangoLayout}} for this scale, or \code{NULL} if the draw_value property
   is \code{FALSE}.
  }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
