\alias{gtkScaleSetDrawValue}
\name{gtkScaleSetDrawValue}
\title{gtkScaleSetDrawValue}
\description{Specifies whether the current value is displayed as a string next to the
slider.}
\usage{gtkScaleSetDrawValue(object, draw.value)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkScale}}] a \verb{\link{GtkScale}}.}
\item{\verb{draw.value}}{[logical] a boolean.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
