\alias{gtkScaleSetDrawValue}
\name{gtkScaleSetDrawValue}
\title{gtkScaleSetDrawValue}
\description{Specifies whether the current value is displayed as a string next 
to the slider.}
\usage{gtkScaleSetDrawValue(object, draw.value)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkScale}}}
\item{\verb{draw.value}}{\code{TRUE} to draw the value}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
