\alias{gtkScaleSetValuePos}
\name{gtkScaleSetValuePos}
\title{gtkScaleSetValuePos}
\description{Sets the position in which the current value is displayed.}
\usage{gtkScaleSetValuePos(object, pos)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkScale}}}
\item{\verb{pos}}{the position in which the current value is displayed}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
