\alias{gtkScrolledWindowGetHadjustment}
\name{gtkScrolledWindowGetHadjustment}
\title{gtkScrolledWindowGetHadjustment}
\description{Returns the horizontal scrollbar's adjustment, used to connect the
horizontal scrollbar to the child widget's horizontal scroll
functionality.}
\usage{gtkScrolledWindowGetHadjustment(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkScrolledWindow}}] A \verb{\link{GtkScrolledWindow}}.}}
\value{[\verb{\link{GtkAdjustment}}] The horizontal \verb{\link{GtkAdjustment}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
