\alias{gtkScrolledWindowGetHadjustment}
\name{gtkScrolledWindowGetHadjustment}
\title{gtkScrolledWindowGetHadjustment}
\description{Returns the horizontal scrollbar's adjustment, used to connect the
horizontal scrollbar to the child widget's horizontal scroll
functionality.}
\usage{gtkScrolledWindowGetHadjustment(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkScrolledWindow}}}}
\value{[\code{\link{GtkAdjustment}}] the horizontal \code{\link{GtkAdjustment}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
