\alias{gtkScrolledWindowGetHscrollbar}
\name{gtkScrolledWindowGetHscrollbar}
\title{gtkScrolledWindowGetHscrollbar}
\description{Returns the horizontal scrollbar of \code{scrolled.window}.}
\usage{gtkScrolledWindowGetHscrollbar(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkScrolledWindow}}}}
\details{Since 2.8}
\value{[\code{\link{GtkWidget}}] the horizontal scrollbar of the scrolled window, or
\code{NULL} if it does not have one.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
