\alias{gtkScrolledWindowGetPlacement}
\name{gtkScrolledWindowGetPlacement}
\title{gtkScrolledWindowGetPlacement}
\description{Gets the placement of the contents with respect to the scrollbars
for the scrolled window. See \code{\link{gtkScrolledWindowSetPlacement}}.}
\usage{gtkScrolledWindowGetPlacement(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkScrolledWindow}}}}
\value{[\code{\link{GtkCornerType}}] the current placement value.
See also \code{\link{gtkScrolledWindowSetPlacement}} and
\code{\link{gtkScrolledWindowUnsetPlacement}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
