\alias{gtkScrolledWindowGetVadjustment}
\name{gtkScrolledWindowGetVadjustment}
\title{gtkScrolledWindowGetVadjustment}
\description{Returns the vertical scrollbar's adjustment, used to connect the
vertical scrollbar to the child widget's vertical scroll
functionality.}
\usage{gtkScrolledWindowGetVadjustment(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkScrolledWindow}}] A \code{\link{GtkScrolledWindow}}.}}
\value{[\code{\link{GtkAdjustment}}] The vertical \code{\link{GtkAdjustment}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
