\alias{gtkScrolledWindowGetVscrollbar}
\name{gtkScrolledWindowGetVscrollbar}
\title{gtkScrolledWindowGetVscrollbar}
\description{Returns the vertical scrollbar of \code{scrolled.window}.}
\usage{gtkScrolledWindowGetVscrollbar(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkScrolledWindow}}}}
\details{Since 2.8}
\value{[\code{\link{GtkWidget}}] the vertical scrollbar of the scrolled window, or
\code{NULL} if it does not have one.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
