\alias{gtkScrolledWindowSetHadjustment}
\name{gtkScrolledWindowSetHadjustment}
\title{gtkScrolledWindowSetHadjustment}
\description{Sets the \verb{\link{GtkAdjustment}} for the horizontal scrollbar.}
\usage{gtkScrolledWindowSetHadjustment(object, hadjustment)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkScrolledWindow}}] A \verb{\link{GtkScrolledWindow}}.}
\item{\verb{hadjustment}}{[\verb{\link{GtkAdjustment}}] Horizontal scroll adjustment.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
