\alias{gtkScrolledWindowSetHadjustment}
\name{gtkScrolledWindowSetHadjustment}
\title{gtkScrolledWindowSetHadjustment}
\description{Sets the \code{\link{GtkAdjustment}} for the horizontal scrollbar.}
\usage{gtkScrolledWindowSetHadjustment(object, hadjustment)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkScrolledWindow}}}
\item{\verb{hadjustment}}{horizontal scroll adjustment}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
