\alias{gtkScrolledWindowSetVadjustment}
\name{gtkScrolledWindowSetVadjustment}
\title{gtkScrolledWindowSetVadjustment}
\description{Sets the \verb{\link{GtkAdjustment}} for the vertical scrollbar.}
\usage{gtkScrolledWindowSetVadjustment(object, hadjustment)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkScrolledWindow}}] A \verb{\link{GtkScrolledWindow}}.}
\item{\verb{hadjustment}}{[\verb{\link{GtkAdjustment}}] Vertical scroll adjustment.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
