\alias{gtkScrolledWindowSetVadjustment}
\name{gtkScrolledWindowSetVadjustment}
\title{gtkScrolledWindowSetVadjustment}
\description{Sets the \code{\link{GtkAdjustment}} for the vertical scrollbar.}
\usage{gtkScrolledWindowSetVadjustment(object, hadjustment)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkScrolledWindow}}}
\item{\verb{hadjustment}}{vertical scroll adjustment}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
