\alias{gtkScrolledWindowUnsetPlacement}
\name{gtkScrolledWindowUnsetPlacement}
\title{gtkScrolledWindowUnsetPlacement}
\description{Unsets the placement of the contents with respect to the scrollbars
for the scrolled window. If no window placement is set for a scrolled
window, it obeys the "gtk-scrolled-window-placement" XSETTING.}
\usage{gtkScrolledWindowUnsetPlacement(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkScrolledWindow}}}}
\details{See also \code{\link{gtkScrolledWindowSetPlacement}} and
\code{\link{gtkScrolledWindowGetPlacement}}.
  Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
