\alias{gtkSelectionAddTarget}
\name{gtkSelectionAddTarget}
\title{gtkSelectionAddTarget}
\description{Appends a specified target to the list of supported targets for a 
given widget and selection.}
\usage{gtkSelectionAddTarget(object, selection, target, info)}
\arguments{
\item{\verb{object}}{a \verb{GtkTarget}}
\item{\verb{selection}}{the selection}
\item{\verb{target}}{target to add.}
\item{\verb{info}}{A unsigned integer which will be passed back to the application.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
