\alias{gtkSelectionAddTargets}
\name{gtkSelectionAddTargets}
\title{gtkSelectionAddTargets}
\description{Prepends a table of targets to the list of supported targets
for a given widget and selection.}
\usage{gtkSelectionAddTargets(object, selection, targets)}
\arguments{
\item{\code{object}}{[\code{\link{GtkWidget}}]  a \code{\link{GtkWidget}}}
\item{\code{selection}}{[\code{\link{GdkAtom}}]  the selection}
\item{\code{targets}}{[\code{\link{GtkTargetEntry}}]  a table of targets to add}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
