\alias{gtkSelectionAddTargets}
\name{gtkSelectionAddTargets}
\title{gtkSelectionAddTargets}
\description{Prepends a table of targets to the list of supported targets
for a given widget and selection.}
\usage{gtkSelectionAddTargets(object, selection, targets)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWidget}}}
\item{\verb{selection}}{the selection}
\item{\verb{targets}}{a table of targets to add}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
