\alias{gtkSelectionClearTargets}
\name{gtkSelectionClearTargets}
\title{gtkSelectionClearTargets}
\description{Remove all targets registered for the given selection for the
widget.}
\usage{gtkSelectionClearTargets(object, selection)}
\arguments{
\item{\code{object}}{[\code{\link{GtkWidget}}]     a \code{\link{GtkWidget}}}
\item{\code{selection}}{[\code{\link{GdkAtom}}]  an atom representing a selection}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
