\alias{gtkSelectionClearTargets}
\name{gtkSelectionClearTargets}
\title{gtkSelectionClearTargets}
\description{Remove all targets registered for the given selection for the
widget.}
\usage{gtkSelectionClearTargets(object, selection)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWidget}}}
\item{\verb{selection}}{an atom representing a selection}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
