\alias{gtkSelectionDataCopy}
\name{gtkSelectionDataCopy}
\title{gtkSelectionDataCopy}
\description{Makes a copy of a \code{\link{GtkSelectionData}} structure and its data.}
\usage{gtkSelectionDataCopy(object)}
\arguments{\item{\verb{object}}{a pointer to a \code{\link{GtkSelectionData}} structure.}}
\value{[\code{\link{GtkSelectionData}}] a pointer to a copy of \code{data}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
