\alias{GtkSeparator}
\name{GtkSeparator}
\title{GtkSeparator}
\description{Base class for GtkHSeparator and GtkVSeparator}
\section{Hierarchy}{\preformatted{GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkSeparator
                           +----GtkHSeparator
                           +----GtkVSeparator}}
\section{Interfaces}{GtkSeparator implements
 AtkImplementorIface and  \verb{\link{GtkBuildable}}.}
\section{Detailed Description}{The \verb{\link{GtkSeparator}} widget is an abstract class, used only for deriving the
subclasses \verb{\link{GtkHSeparator}} and \verb{\link{GtkVSeparator}}.}
\section{Structures}{\describe{\item{\verb{GtkSeparator}}{
The \verb{\link{GtkSeparator}} struct contains private data only.

}}}
\references{\url{http://developer.gnome.org/doc/API/2.0/gtk/GtkSeparator.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
