\alias{GtkSeparator}
\name{GtkSeparator}
\title{GtkSeparator}
\description{Base class for GtkHSeparator and GtkVSeparator}
\section{Hierarchy}{\preformatted{  \link{GObject}
   +----GInitiallyUnowned
         +----\link{GtkObject}
               +----\link{GtkWidget}
                     +----GtkSeparator
                           +----\link{GtkHSeparator}
                           +----\link{GtkVSeparator}}}
\section{Interfaces}{GtkSeparator implements
 \code{\link{GtkBuildable}} and  AtkImplementorIface.}
\section{Detailed Description}{The \code{\link{GtkSeparator}} widget is an abstract class, used only for deriving the
subclasses \code{\link{GtkHSeparator}} and \code{\link{GtkVSeparator}}.}
\section{Structures}{\describe{\item{\code{GtkSeparator}}{
The \code{\link{GtkSeparator}} struct contains private data only.

}}}
\references{\url{http://developer.gnome.org/doc/API/2.0/gtk/GtkSeparator.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
