\alias{GtkSeparator}
\name{GtkSeparator}
\title{GtkSeparator}
\description{Base class for GtkHSeparator and GtkVSeparator}
\section{Hierarchy}{\preformatted{GObject
   +----GInitiallyUnowned
         +----GtkObject
               +----GtkWidget
                     +----GtkSeparator
                           +----GtkHSeparator
                           +----GtkVSeparator}}
\section{Interfaces}{GtkSeparator implements
 AtkImplementorIface,  \code{\link{GtkBuildable}} and  \code{\link{GtkOrientable}}.}
\section{Detailed Description}{The \code{\link{GtkSeparator}} widget is an abstract class, used only for deriving the
subclasses \code{\link{GtkHSeparator}} and \code{\link{GtkVSeparator}}.}
\section{Structures}{\describe{\item{\verb{GtkSeparator}}{
The \code{\link{GtkSeparator}} struct contains private data only.

}}}
\references{\url{https://developer-old.gnome.org/gtk2/stable/GtkSeparator.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
