\alias{gtkSeparatorMenuItemNew}
\name{gtkSeparatorMenuItemNew}
\title{gtkSeparatorMenuItemNew}
\description{Creates a new \verb{\link{GtkSeparatorMenuItem}}.}
\usage{gtkSeparatorMenuItemNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}] a new \verb{\link{GtkSeparatorMenuItem}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
