\alias{gtkSeparatorToolItemGetDraw}
\name{gtkSeparatorToolItemGetDraw}
\title{gtkSeparatorToolItemGetDraw}
\description{Returns whether \code{item} is drawn as a line, or just blank. 
See \code{\link{gtkSeparatorToolItemSetDraw}}.}
\usage{gtkSeparatorToolItemGetDraw(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkSeparatorToolItem}}}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if \code{item} is drawn as a line, or just blank.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
