\alias{gtkSeparatorToolItemNew}
\name{gtkSeparatorToolItemNew}
\title{gtkSeparatorToolItemNew}
\description{Create a new \code{\link{GtkSeparatorToolItem}}}
\usage{gtkSeparatorToolItemNew(show = TRUE)}
\details{Since 2.4}
\value{[\code{\link{GtkToolItem}}] the new \code{\link{GtkSeparatorToolItem}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
