\alias{gtkSeparatorToolItemSetDraw}
\name{gtkSeparatorToolItemSetDraw}
\title{gtkSeparatorToolItemSetDraw}
\description{Whether \code{item} is drawn as a vertical line, or just blank.
Setting this to \code{FALSE} along with \code{\link{gtkToolItemSetExpand}} is useful
to create an item that forces following items to the end of the toolbar.}
\usage{gtkSeparatorToolItemSetDraw(object, draw)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkSeparatorToolItem}}}
\item{\verb{draw}}{whether \code{item} is drawn as a vertical line}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
