\alias{gtkSettingsGetDefault}
\name{gtkSettingsGetDefault}
\title{gtkSettingsGetDefault}
\description{Gets the \code{\link{GtkSettings}} object for the default GDK screen, creating
it if necessary. See \code{\link{gtkSettingsGetForScreen}}.}
\usage{gtkSettingsGetDefault()}
\value{[\code{\link{GtkSettings}}] a \code{\link{GtkSettings}} object. If there is no default
screen, then returns \code{NULL}. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
