\alias{gtkSettingsGetForScreen}
\name{gtkSettingsGetForScreen}
\title{gtkSettingsGetForScreen}
\description{Gets the \code{\link{GtkSettings}} object for \code{screen}, creating it if necessary.}
\usage{gtkSettingsGetForScreen(screen)}
\arguments{\item{\verb{screen}}{a \code{\link{GdkScreen}}.}}
\details{Since 2.2}
\value{[\code{\link{GtkSettings}}] a \code{\link{GtkSettings}} object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
