\alias{gtkSettingsInstallProperty}
\name{gtkSettingsInstallProperty}
\title{gtkSettingsInstallProperty}
\description{\emph{undocumented
}}
\usage{gtkSettingsInstallProperty(pspec)}
\arguments{\item{\verb{pspec}}{[\verb{\link{GParamSpec}}] }}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
