\alias{gtkSettingsInstallProperty}
\name{gtkSettingsInstallProperty}
\title{gtkSettingsInstallProperty}
\description{\emph{undocumented
}}
\usage{gtkSettingsInstallProperty(pspec)}
\arguments{\item{\code{pspec}}{[\code{\link{GParamSpec}}] }}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
