\alias{gtkSettingsInstallProperty}
\name{gtkSettingsInstallProperty}
\title{gtkSettingsInstallProperty}
\description{\emph{undocumented
}}
\usage{gtkSettingsInstallProperty(pspec)}
\arguments{\item{\verb{pspec}}{\emph{undocumented
}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
