\alias{gtkSettingsInstallPropertyParser}
\name{gtkSettingsInstallPropertyParser}
\title{gtkSettingsInstallPropertyParser}
\description{\emph{undocumented
}}
\usage{gtkSettingsInstallPropertyParser(pspec, parser)}
\arguments{
\item{\verb{pspec}}{[\verb{\link{GParamSpec}}] }
\item{\verb{parser}}{[GtkRcPropertyParser] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
