\alias{gtkSettingsInstallPropertyParser}
\name{gtkSettingsInstallPropertyParser}
\title{gtkSettingsInstallPropertyParser}
\usage{gtkSettingsInstallPropertyParser(pspec, parser)}
\arguments{
\item{\code{pspec}}{[\code{\link{GParamSpec}}] }
\item{\code{parser}}{[GtkRcPropertyParser] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
