\alias{gtkSettingsInstallPropertyParser}
\name{gtkSettingsInstallPropertyParser}
\title{gtkSettingsInstallPropertyParser}
\description{\emph{undocumented
}}
\usage{gtkSettingsInstallPropertyParser(pspec, parser)}
\arguments{
\item{\verb{pspec}}{\emph{undocumented
}}
\item{\verb{parser}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
