\alias{gtkSettingsSetDoubleProperty}
\name{gtkSettingsSetDoubleProperty}
\title{gtkSettingsSetDoubleProperty}
\description{\emph{undocumented
}}
\usage{gtkSettingsSetDoubleProperty(object, name, v.double, origin)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkSettings}}] }
\item{\verb{name}}{[character] }
\item{\verb{v.double}}{[numeric] }
\item{\verb{origin}}{[character] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
