\alias{gtkSettingsSetDoubleProperty}
\name{gtkSettingsSetDoubleProperty}
\title{gtkSettingsSetDoubleProperty}
\description{\emph{undocumented
}}
\usage{gtkSettingsSetDoubleProperty(object, name, v.double, origin)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{name}}{\emph{undocumented
}}
\item{\verb{v.double}}{\emph{undocumented
}}
\item{\verb{origin}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
