\alias{gtkSettingsSetLongProperty}
\name{gtkSettingsSetLongProperty}
\title{gtkSettingsSetLongProperty}
\description{\emph{undocumented
}}
\usage{gtkSettingsSetLongProperty(object, name, v.long, origin)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkSettings}}] }
\item{\verb{name}}{[character] }
\item{\verb{v.long}}{[numeric] }
\item{\verb{origin}}{[character] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
