\alias{gtkSettingsSetLongProperty}
\name{gtkSettingsSetLongProperty}
\title{gtkSettingsSetLongProperty}
\description{\emph{undocumented
}}
\usage{gtkSettingsSetLongProperty(object, name, v.long, origin)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{name}}{\emph{undocumented
}}
\item{\verb{v.long}}{\emph{undocumented
}}
\item{\verb{origin}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
