\alias{gtkSettingsSetPropertyValue}
\name{gtkSettingsSetPropertyValue}
\title{gtkSettingsSetPropertyValue}
\description{\emph{undocumented
}}
\usage{gtkSettingsSetPropertyValue(object, name, svalue)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkSettings}}] }
\item{\verb{name}}{[character] }
\item{\verb{svalue}}{[\verb{\link{GtkSettingsValue}}] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
