\alias{gtkSettingsSetPropertyValue}
\name{gtkSettingsSetPropertyValue}
\title{gtkSettingsSetPropertyValue}
\description{\emph{undocumented
}}
\usage{gtkSettingsSetPropertyValue(object, name, svalue)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{name}}{\emph{undocumented
}}
\item{\verb{svalue}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
