\alias{gtkSettingsSetStringProperty}
\name{gtkSettingsSetStringProperty}
\title{gtkSettingsSetStringProperty}
\usage{gtkSettingsSetStringProperty(object, name, v.string, origin)}
\arguments{
\item{\code{object}}{[\code{\link{GtkSettings}}] }
\item{\code{name}}{[character] }
\item{\code{v.string}}{[character] }
\item{\code{origin}}{[character] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
