\alias{gtkSettingsSetStringProperty}
\name{gtkSettingsSetStringProperty}
\title{gtkSettingsSetStringProperty}
\description{\emph{undocumented
}}
\usage{gtkSettingsSetStringProperty(object, name, v.string, origin)}
\arguments{
\item{\verb{object}}{\emph{undocumented
}}
\item{\verb{name}}{\emph{undocumented
}}
\item{\verb{v.string}}{\emph{undocumented
}}
\item{\verb{origin}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
