\alias{gtkShowAboutDialog}
\name{gtkShowAboutDialog}
\title{gtkShowAboutDialog}
\description{This is a convenience function for showing an application's about box.
The constructed dialog is associated with the parent window and 
reused for future invocations of this function.}
\usage{gtkShowAboutDialog(parent, ...)}
\arguments{
\item{\code{parent}}{[\code{\link{GtkWindow}}]  transient parent, or \code{NULL} for none}
\item{\code{...}}{ value of first property, followed by more properties, \code{NULL}-terminated}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
