\alias{gtkSizeGroupGetIgnoreHidden}
\name{gtkSizeGroupGetIgnoreHidden}
\title{gtkSizeGroupGetIgnoreHidden}
\description{Returns if invisible widgets are ignored when calculating the size.}
\usage{gtkSizeGroupGetIgnoreHidden(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkSizeGroup}}}}
\details{Since 2.8}
\value{[logical] \code{TRUE} if invisible widgets are ignored.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
