\alias{gtkSizeGroupGetWidgets}
\name{gtkSizeGroupGetWidgets}
\title{gtkSizeGroupGetWidgets}
\description{Returns the list of widgets associated with \code{size.group}.}
\usage{gtkSizeGroupGetWidgets(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkSizeGroup}}]  a \verb{GtkSizeGrup}}}
\details{  Since  2.10}
\value{[list]  a \verb{list} of widgets.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
