\alias{gtkSizeGroupGetWidgets}
\name{gtkSizeGroupGetWidgets}
\title{gtkSizeGroupGetWidgets}
\description{Returns the list of widgets associated with \code{size.group}.}
\usage{gtkSizeGroupGetWidgets(object)}
\arguments{\item{\verb{object}}{a \verb{GtkSizeGrup}}}
\details{Since 2.10}
\value{[list] a \verb{list} of
widgets. \emph{[  \acronym{element-type}   GtkWidget][  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
