\alias{gtkSizeGroupNew}
\name{gtkSizeGroupNew}
\title{gtkSizeGroupNew}
\description{Create a new \code{\link{GtkSizeGroup}}.}
\usage{gtkSizeGroupNew(mode = NULL)}
\arguments{\item{\code{mode}}{[\code{\link{GtkSizeGroupMode}}]  the mode for the new size group.}}
\value{[\code{\link{GtkSizeGroup}}]  a newly created \code{\link{GtkSizeGroup}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
