\alias{gtkSizeGroupNew}
\name{gtkSizeGroupNew}
\title{gtkSizeGroupNew}
\description{Create a new \code{\link{GtkSizeGroup}}.}
\usage{gtkSizeGroupNew(mode = NULL)}
\arguments{\item{\verb{mode}}{the mode for the new size group.}}
\value{[\code{\link{GtkSizeGroup}}] a newly created \code{\link{GtkSizeGroup}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
