\alias{gtkSizeGroupRemoveWidget}
\name{gtkSizeGroupRemoveWidget}
\title{gtkSizeGroupRemoveWidget}
\description{Removes a widget from a \code{\link{GtkSizeGroup}}.}
\usage{gtkSizeGroupRemoveWidget(object, widget)}
\arguments{
\item{\code{object}}{[\code{\link{GtkSizeGroup}}]  a \code{GtkSizeGrup}}
\item{\code{widget}}{[\code{\link{GtkWidget}}]  the \code{\link{GtkWidget}} to remove}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
