\alias{gtkSizeGroupRemoveWidget}
\name{gtkSizeGroupRemoveWidget}
\title{gtkSizeGroupRemoveWidget}
\description{Removes a widget from a \code{\link{GtkSizeGroup}}.}
\usage{gtkSizeGroupRemoveWidget(object, widget)}
\arguments{
\item{\verb{object}}{a \verb{GtkSizeGrup}}
\item{\verb{widget}}{the \code{\link{GtkWidget}} to remove}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
