\alias{gtkSizeGroupSetIgnoreHidden}
\name{gtkSizeGroupSetIgnoreHidden}
\title{gtkSizeGroupSetIgnoreHidden}
\description{Sets whether unmapped widgets should be ignored when
calculating the size.}
\usage{gtkSizeGroupSetIgnoreHidden(object, ignore.hidden)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkSizeGroup}}}
\item{\verb{ignore.hidden}}{whether unmapped widgets should be ignored
when calculating the size}
}
\details{Since 2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
