\alias{gtkSpinButtonGetAdjustment}
\name{gtkSpinButtonGetAdjustment}
\title{gtkSpinButtonGetAdjustment}
\description{Get the adjustment associated with a \code{\link{GtkSpinButton}}}
\usage{gtkSpinButtonGetAdjustment(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkSpinButton}}}}
\value{[\code{\link{GtkAdjustment}}] the \code{\link{GtkAdjustment}} of \code{spin.button}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
