\alias{gtkSpinButtonGetRange}
\name{gtkSpinButtonGetRange}
\title{gtkSpinButtonGetRange}
\description{Gets the range allowed for \code{spin.button}. See
\code{\link{gtkSpinButtonSetRange}}.}
\usage{gtkSpinButtonGetRange(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkSpinButton}}]  a \code{\link{GtkSpinButton}}}}
\value{
A list containing the following elements:
\item{\code{min}}{[numeric]  location to store minimum allowed value, or \code{NULL}}
\item{\code{max}}{[numeric]  location to store maximum allowed value, or \code{NULL}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
