\alias{gtkSpinButtonGetValueAsInt}
\name{gtkSpinButtonGetValueAsInt}
\title{gtkSpinButtonGetValueAsInt}
\description{Get the value \code{spin.button} represented as an integer.}
\usage{gtkSpinButtonGetValueAsInt(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkSpinButton}}}}
\value{[integer] the value of \code{spin.button}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
