\alias{gtkSpinButtonSetAdjustment}
\name{gtkSpinButtonSetAdjustment}
\title{gtkSpinButtonSetAdjustment}
\description{Replaces the \code{\link{GtkAdjustment}} associated with \code{spin.button}.}
\usage{gtkSpinButtonSetAdjustment(object, adjustment)}
\arguments{
\item{\code{object}}{[\code{\link{GtkSpinButton}}]  a \code{\link{GtkSpinButton}}}
\item{\code{adjustment}}{[\code{\link{GtkAdjustment}}]  a \code{\link{GtkAdjustment}} to replace the existing adjustment}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
