\alias{gtkSpinButtonSetIncrements}
\name{gtkSpinButtonSetIncrements}
\title{gtkSpinButtonSetIncrements}
\description{Sets the step and page increments for spin_button.  This affects how 
quickly the value changes when the spin button's arrows are activated.}
\usage{gtkSpinButtonSetIncrements(object, step, page)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkSpinButton}}}
\item{\verb{step}}{increment applied for a button 1 press.}
\item{\verb{page}}{increment applied for a button 2 press.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
