\alias{gtkSpinButtonSetRange}
\name{gtkSpinButtonSetRange}
\title{gtkSpinButtonSetRange}
\description{Sets the minimum and maximum allowable values for \code{spin.button}}
\usage{gtkSpinButtonSetRange(object, min, max)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkSpinButton}}]  a \verb{\link{GtkSpinButton}}}
\item{\verb{min}}{[numeric]  minimum allowable value}
\item{\verb{max}}{[numeric]  maximum allowable value}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
