\alias{gtkSpinButtonSetUpdatePolicy}
\name{gtkSpinButtonSetUpdatePolicy}
\title{gtkSpinButtonSetUpdatePolicy}
\description{Sets the update behavior of a spin button. This determines whether the
spin button is always updated or only when a valid value is set.}
\usage{gtkSpinButtonSetUpdatePolicy(object, policy)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkSpinButton}}}
\item{\verb{policy}}{a \code{\link{GtkSpinButtonUpdatePolicy}} value}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
