\alias{gtkSpinButtonSetValue}
\name{gtkSpinButtonSetValue}
\title{gtkSpinButtonSetValue}
\description{Set the value of \code{spin.button}.}
\usage{gtkSpinButtonSetValue(object, value)}
\arguments{
\item{\code{object}}{[\code{\link{GtkSpinButton}}]  a \code{\link{GtkSpinButton}}}
\item{\code{value}}{[numeric]  the new value}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
