\alias{gtkSpinButtonSetValue}
\name{gtkSpinButtonSetValue}
\title{gtkSpinButtonSetValue}
\description{Set the value of \code{spin.button}.}
\usage{gtkSpinButtonSetValue(object, value)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkSpinButton}}}
\item{\verb{value}}{the new value}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
