\alias{gtkSpinButtonUpdate}
\name{gtkSpinButtonUpdate}
\title{gtkSpinButtonUpdate}
\description{Manually force an update of the spin button.}
\usage{gtkSpinButtonUpdate(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkSpinButton}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
