\alias{gtkStatusbarGetHasResizeGrip}
\name{gtkStatusbarGetHasResizeGrip}
\title{gtkStatusbarGetHasResizeGrip}
\description{Returns whether the statusbar has a resize grip.}
\usage{gtkStatusbarGetHasResizeGrip(object)}
\arguments{\item{\verb{object}}{a \verb{GtkStatusBar}}}
\value{[logical] \code{TRUE} if the statusbar has a resize grip.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
