\alias{gtkStatusbarNew}
\name{gtkStatusbarNew}
\title{gtkStatusbarNew}
\description{Creates a new \verb{\link{GtkStatusbar}} ready for messages.}
\usage{gtkStatusbarNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}]  the new \verb{\link{GtkStatusbar}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
