\alias{gtkStatusbarPop}
\name{gtkStatusbarPop}
\title{gtkStatusbarPop}
\description{Removes the message at the top of a \code{GtkStatusBar}'s stack.}
\usage{gtkStatusbarPop(object, context.id)}
\arguments{
\item{\code{object}}{[\code{\link{GtkStatusbar}}] a \code{GtkStatusBar}.}
\item{\code{context.id}}{[numeric] a context identifier.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
