\alias{gtkStatusbarSetHasResizeGrip}
\name{gtkStatusbarSetHasResizeGrip}
\title{gtkStatusbarSetHasResizeGrip}
\description{Sets whether the statusbar has a resize grip. 
\code{TRUE} by default.}
\usage{gtkStatusbarSetHasResizeGrip(object, setting)}
\arguments{
\item{\code{object}}{[\code{\link{GtkStatusbar}}]  a \code{GtkStatusBar}}
\item{\code{setting}}{[logical] \code{TRUE} to have a resize grip}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
