\alias{gtkStatusbarSetHasResizeGrip}
\name{gtkStatusbarSetHasResizeGrip}
\title{gtkStatusbarSetHasResizeGrip}
\description{Sets whether the statusbar has a resize grip. 
\code{TRUE} by default.}
\usage{gtkStatusbarSetHasResizeGrip(object, setting)}
\arguments{
\item{\verb{object}}{a \verb{GtkStatusBar}}
\item{\verb{setting}}{\code{TRUE} to have a resize grip}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
