\alias{gtkStatusIconGetScreen}
\name{gtkStatusIconGetScreen}
\title{gtkStatusIconGetScreen}
\description{Returns the \code{\link{GdkScreen}} associated with \code{status.icon}.}
\usage{gtkStatusIconGetScreen(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkStatusIcon}}}}
\details{Since 2.12}
\value{[\code{\link{GdkScreen}}] a \code{\link{GdkScreen}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
