\alias{gtkStatusIconGetStorageType}
\name{gtkStatusIconGetStorageType}
\title{gtkStatusIconGetStorageType}
\description{Gets the type of representation being used by the \code{\link{GtkStatusIcon}}
to store image data. If the \code{\link{GtkStatusIcon}} has no image data,
the return value will be \code{GTK_IMAGE_EMPTY}.}
\usage{gtkStatusIconGetStorageType(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkStatusIcon}}}}
\details{Since 2.10}
\value{[\code{\link{GtkImageType}}] the image representation being used}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
