\alias{gtkStatusIconNew}
\name{gtkStatusIconNew}
\title{gtkStatusIconNew}
\description{Creates an empty status icon object.}
\usage{gtkStatusIconNew()}
\details{Since 2.10}
\value{[\code{\link{GtkStatusIcon}}] a new \code{\link{GtkStatusIcon}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
