\alias{gtkStatusIconSetFromFile}
\name{gtkStatusIconSetFromFile}
\title{gtkStatusIconSetFromFile}
\description{Makes \code{status.icon} display the file \code{filename}.
See \code{\link{gtkStatusIconNewFromFile}} for details.}
\usage{gtkStatusIconSetFromFile(object, filename)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkStatusIcon}}}
\item{\verb{filename}}{a filename}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
