\alias{gtkStatusIconSetFromPixbuf}
\name{gtkStatusIconSetFromPixbuf}
\title{gtkStatusIconSetFromPixbuf}
\description{Makes \code{status.icon} display \code{pixbuf}. 
See \code{\link{gtkStatusIconNewFromPixbuf}} for details.}
\usage{gtkStatusIconSetFromPixbuf(object, pixbuf)}
\arguments{
\item{\code{object}}{[\code{\link{GtkStatusIcon}}]  a \code{\link{GtkStatusIcon}}}
\item{\code{pixbuf}}{[\code{\link{GdkPixbuf}}]  a \code{\link{GdkPixbuf}} or \code{NULL}}
}
\details{  Since  2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
