\alias{gtkStatusIconSetFromPixbuf}
\name{gtkStatusIconSetFromPixbuf}
\title{gtkStatusIconSetFromPixbuf}
\description{Makes \code{status.icon} display \code{pixbuf}.
See \code{\link{gtkStatusIconNewFromPixbuf}} for details.}
\usage{gtkStatusIconSetFromPixbuf(object, pixbuf)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkStatusIcon}}}
\item{\verb{pixbuf}}{a \code{\link{GdkPixbuf}} or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
