\alias{gtkStatusIconSetScreen}
\name{gtkStatusIconSetScreen}
\title{gtkStatusIconSetScreen}
\description{Sets the \verb{\link{GdkScreen}} where \code{status.icon} is displayed; if
the icon is already mapped, it will be unmapped, and
then remapped on the new screen.}
\usage{gtkStatusIconSetScreen(object, screen)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkStatusIcon}}]  a \verb{\link{GtkStatusIcon}}}
\item{\verb{screen}}{[\verb{\link{GdkScreen}}]  a \verb{\link{GdkScreen}}}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
