\alias{gtkStatusIconSetScreen}
\name{gtkStatusIconSetScreen}
\title{gtkStatusIconSetScreen}
\description{Sets the \code{\link{GdkScreen}} where \code{status.icon} is displayed; if
the icon is already mapped, it will be unmapped, and
then remapped on the new screen.}
\usage{gtkStatusIconSetScreen(object, screen)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkStatusIcon}}}
\item{\verb{screen}}{a \code{\link{GdkScreen}}}
}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
