\alias{gtkStatusIconSetTooltip}
\name{gtkStatusIconSetTooltip}
\title{gtkStatusIconSetTooltip}
\description{
Sets the tooltip of the status icon.
\strong{WARNING: \code{gtk_status_icon_set_tooltip} has been deprecated since version 2.16 and should not be used in newly-written code. Use \code{\link{gtkStatusIconSetTooltipText}} instead.}
}
\usage{gtkStatusIconSetTooltip(object, tooltip.text)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkStatusIcon}}}
\item{\verb{tooltip.text}}{the tooltip text, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
