\alias{gtkStatusIconSetVisible}
\name{gtkStatusIconSetVisible}
\title{gtkStatusIconSetVisible}
\description{Shows or hides a status icon.}
\usage{gtkStatusIconSetVisible(object, visible)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkStatusIcon}}}
\item{\verb{visible}}{\code{TRUE} to show the status icon, \code{FALSE} to hide it}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
