\alias{gtkStockItemCopy}
\name{gtkStockItemCopy}
\title{gtkStockItemCopy}
\description{Copies a stock item, mostly useful for language bindings and not in applications.}
\usage{gtkStockItemCopy(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkStockItem}}}}
\value{[\code{\link{GtkStockItem}}] a new \code{\link{GtkStockItem}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
