\alias{gtkStockListIds}
\name{gtkStockListIds}
\title{gtkStockListIds}
\description{Retrieves a list of all known stock IDs added to a \code{\link{GtkIconFactory}}
or registered with \code{\link{gtkStockAdd}}.}
\usage{gtkStockListIds()}
\value{[list] a list of known stock IDs. \emph{[  \acronym{element-type}   utf8][  \acronym{transfer full}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
