\alias{gtkStyleApplyDefaultBackground}
\name{gtkStyleApplyDefaultBackground}
\title{gtkStyleApplyDefaultBackground}
\description{\emph{undocumented
}}
\usage{gtkStyleApplyDefaultBackground(object, window, set.bg, state.type, area = NULL, x, y, width, height)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkStyle}}] }
\item{\verb{window}}{[\verb{\link{GdkWindow}}] }
\item{\verb{set.bg}}{[logical] }
\item{\verb{state.type}}{[\code{\link{GtkStateType}}] }
\item{\verb{area}}{[\verb{\link{GdkRectangle}}] }
\item{\verb{x}}{[integer] }
\item{\verb{y}}{[integer] }
\item{\verb{width}}{[integer] }
\item{\verb{height}}{[integer] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
