\alias{gtkStyleApplyDefaultBackground}
\name{gtkStyleApplyDefaultBackground}
\title{gtkStyleApplyDefaultBackground}
\description{\emph{undocumented
}}
\usage{gtkStyleApplyDefaultBackground(object, window, set.bg, state.type, 
    area = NULL, x, y, width, height)}
\arguments{
\item{\verb{object}}{. \emph{[  \acronym{allow-none}  ]}}
\item{\verb{window}}{\emph{undocumented
}}
\item{\verb{set.bg}}{\emph{undocumented
}}
\item{\verb{state.type}}{\emph{undocumented
}}
\item{\verb{area}}{\emph{undocumented
}}
\item{\verb{x}}{\emph{undocumented
}}
\item{\verb{y}}{\emph{undocumented
}}
\item{\verb{width}}{\emph{undocumented
}}
\item{\verb{height}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
