\alias{gtkStyleLookupIconSet}
\name{gtkStyleLookupIconSet}
\title{gtkStyleLookupIconSet}
\description{\emph{undocumented
}}
\usage{gtkStyleLookupIconSet(object, stock.id)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkStyle}}] }
\item{\verb{stock.id}}{[character] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
