\alias{gtkStyleLookupIconSet}
\name{gtkStyleLookupIconSet}
\title{gtkStyleLookupIconSet}
\description{\emph{undocumented
}}
\usage{gtkStyleLookupIconSet(object, stock.id)}
\arguments{
\item{\code{object}}{[\code{\link{GtkStyle}}] }
\item{\code{stock.id}}{[character] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
