\alias{gtkStyleSetBackground}
\name{gtkStyleSetBackground}
\title{gtkStyleSetBackground}
\description{Sets the background of \code{window} to the background color or pixmap
specified by \code{style} for the given state.}
\usage{gtkStyleSetBackground(object, window, state.type)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkStyle}}}
\item{\verb{window}}{a \code{\link{GdkWindow}}}
\item{\verb{state.type}}{a state}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
