\alias{gtkTableGetHomogeneous}
\name{gtkTableGetHomogeneous}
\title{gtkTableGetHomogeneous}
\description{Returns whether the table cells are all constrained to the same
width and height. (See \code{gtkTableSetHomogenous()})}
\usage{gtkTableGetHomogeneous(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTable}}]  a \code{\link{GtkTable}}}}
\value{[logical] \code{TRUE} if the cells are all constrained to the same size}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
