\alias{gtkTableGetRowSpacing}
\name{gtkTableGetRowSpacing}
\title{gtkTableGetRowSpacing}
\description{Gets the amount of space between row \code{row}, and
row \code{row} + 1. See \code{\link{gtkTableSetRowSpacing}}.}
\usage{gtkTableGetRowSpacing(object, row)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTable}}}
\item{\verb{row}}{a row in the table, 0 indicates the first row}
}
\value{[numeric] the row spacing}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
