\alias{gtkTableResize}
\name{gtkTableResize}
\title{gtkTableResize}
\description{If you need to change a table's size \emph{after} it has been created, this function allows you to do so.}
\usage{gtkTableResize(object, rows, columns)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTable}}] The \code{\link{GtkTable}} you wish to change the size of.}
\item{\code{rows}}{[numeric] The new number of rows.}
\item{\code{columns}}{[numeric] The new number of columns.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
