\alias{gtkTableSetColSpacing}
\name{gtkTableSetColSpacing}
\title{gtkTableSetColSpacing}
\description{Alters the amount of space between a given table column and the adjacent columns.}
\usage{gtkTableSetColSpacing(object, column, spacing)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTable}}] a \code{\link{GtkTable}}.}
\item{\code{column}}{[numeric] the column whose spacing should be changed.}
\item{\code{spacing}}{[numeric] number of pixels that the spacing should take up.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
