\alias{gtkTableSetColSpacings}
\name{gtkTableSetColSpacings}
\title{gtkTableSetColSpacings}
\description{Sets the space between every column in \code{table} equal to \code{spacing}.}
\usage{gtkTableSetColSpacings(object, spacing)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTable}}.}
\item{\verb{spacing}}{the number of pixels of space to place between every column in the table.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
