\alias{gtkTableSetRowSpacing}
\name{gtkTableSetRowSpacing}
\title{gtkTableSetRowSpacing}
\description{Changes the space between a given table row and the subsequent row.}
\usage{gtkTableSetRowSpacing(object, row, spacing)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTable}} containing the row whose properties you wish to change.}
\item{\verb{row}}{row number whose spacing will be changed.}
\item{\verb{spacing}}{number of pixels that the spacing should take up.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
