\alias{gtkTargetListAdd}
\name{gtkTargetListAdd}
\title{gtkTargetListAdd}
\description{Appends another target to a \code{\link{GtkTargetList}}.}
\usage{gtkTargetListAdd(object, target, flags, info)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTargetList}}]   a \code{\link{GtkTargetList}}}
\item{\code{target}}{[\code{\link{GdkAtom}}]  the interned atom representing the target}
\item{\code{flags}}{[numeric]  the flags for this target}
\item{\code{info}}{[numeric]  an ID that will be passed back to the application}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
