\alias{gtkTargetListAdd}
\name{gtkTargetListAdd}
\title{gtkTargetListAdd}
\description{Appends another target to a \code{\link{GtkTargetList}}.}
\usage{gtkTargetListAdd(object, target, flags, info)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTargetList}}}
\item{\verb{target}}{the interned atom representing the target}
\item{\verb{flags}}{the flags for this target}
\item{\verb{info}}{an ID that will be passed back to the application}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
