\alias{gtkTargetListAddTable}
\name{gtkTargetListAddTable}
\title{gtkTargetListAddTable}
\description{Prepends a table of \code{\link{GtkTargetEntry}} to a target list.}
\usage{gtkTargetListAddTable(object, targets)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTargetList}}]  a \code{\link{GtkTargetList}}}
\item{\code{targets}}{[\code{\link{GtkTargetEntry}}]  the table of \code{\link{GtkTargetEntry}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
