\alias{gtkTargetListAddTable}
\name{gtkTargetListAddTable}
\title{gtkTargetListAddTable}
\description{Prepends a table of \code{\link{GtkTargetEntry}} to a target list.}
\usage{gtkTargetListAddTable(object, targets)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTargetList}}}
\item{\verb{targets}}{the table of \code{\link{GtkTargetEntry}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
