\alias{gtkTargetListAddTextTargets}
\name{gtkTargetListAddTextTargets}
\title{gtkTargetListAddTextTargets}
\description{Appends the text targets supported by \verb{GtkSelection} to
the target list. All targets are added with the same \code{info}.}
\usage{gtkTargetListAddTextTargets(list, info)}
\arguments{
\item{\verb{list}}{a \code{\link{GtkTargetList}}}
\item{\verb{info}}{an ID that will be passed back to the application}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
