\alias{gtkTargetListAddUriTargets}
\name{gtkTargetListAddUriTargets}
\title{gtkTargetListAddUriTargets}
\description{Appends the URI targets supported by \verb{GtkSelection} to
the target list. All targets are added with the same \code{info}.}
\usage{gtkTargetListAddUriTargets(list, info)}
\arguments{
\item{\verb{list}}{a \code{\link{GtkTargetList}}}
\item{\verb{info}}{an ID that will be passed back to the application}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
