\alias{gtkTargetListNew}
\name{gtkTargetListNew}
\title{gtkTargetListNew}
\description{Creates a new \verb{\link{GtkTargetList}} from a list of \verb{\link{GtkTargetEntry}}.}
\usage{gtkTargetListNew(targets)}
\arguments{\item{\verb{targets}}{[\verb{\link{GtkTargetEntry}}]  Pointer to a list of \verb{\link{GtkTargetEntry}}}}
\value{[\verb{\link{GtkTargetList}}]  the new \verb{\link{GtkTargetList}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
