\alias{gtkTargetListNew}
\name{gtkTargetListNew}
\title{gtkTargetListNew}
\description{Creates a new \code{\link{GtkTargetList}} from a list of \code{\link{GtkTargetEntry}}.}
\usage{gtkTargetListNew(targets)}
\arguments{\item{\verb{targets}}{Pointer to a list of \code{\link{GtkTargetEntry}}}}
\value{[\code{\link{GtkTargetList}}] the new \code{\link{GtkTargetList}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
