\alias{gtkTargetListRemove}
\name{gtkTargetListRemove}
\title{gtkTargetListRemove}
\description{Removes a target from a target list.}
\usage{gtkTargetListRemove(object, target)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTargetList}}]  a \code{\link{GtkTargetList}}}
\item{\code{target}}{[\code{\link{GdkAtom}}]  the interned atom representing the target}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
