\alias{gtkTargetListRemove}
\name{gtkTargetListRemove}
\title{gtkTargetListRemove}
\description{Removes a target from a target list.}
\usage{gtkTargetListRemove(object, target)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTargetList}}}
\item{\verb{target}}{the interned atom representing the target}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
