\alias{gtkTargetsIncludeUri}
\name{gtkTargetsIncludeUri}
\title{gtkTargetsIncludeUri}
\description{Determines if any of the targets in \code{targets} can be used to
provide an uri list.}
\usage{gtkTargetsIncludeUri(targets)}
\arguments{\item{\verb{targets}}{a list of \code{\link{GdkAtom}}s}}
\details{Since 2.10}
\value{[logical] \code{TRUE} if \code{targets} include a suitable target for uri lists,
otherwise \code{FALSE}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
