\alias{gtkTextAttributesCopy}
\name{gtkTextAttributesCopy}
\title{gtkTextAttributesCopy}
\description{Copies \code{src} and returns a new \code{\link{GtkTextAttributes}}.}
\usage{gtkTextAttributesCopy(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextAttributes}} to be copied}}
\value{[\code{\link{GtkTextAttributes}}] a copy of \code{src}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
