\alias{gtkTextAttributesCopyValues}
\name{gtkTextAttributesCopyValues}
\title{gtkTextAttributesCopyValues}
\description{Copies the values from \code{src} to \code{dest} so that \code{dest} has the same values
as \code{src}. Frees existing values in \code{dest}.}
\usage{gtkTextAttributesCopyValues(object, dest)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextAttributes}}}
\item{\verb{dest}}{another \code{\link{GtkTextAttributes}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
