\alias{gtkTextAttributesNew}
\name{gtkTextAttributesNew}
\title{gtkTextAttributesNew}
\description{Creates a \verb{\link{GtkTextAttributes}}, which describes
a set of properties on some text.}
\usage{gtkTextAttributesNew()}
\value{[\verb{\link{GtkTextAttributes}}]  a new \verb{\link{GtkTextAttributes}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
