\alias{gtkTextAttributesNew}
\name{gtkTextAttributesNew}
\title{gtkTextAttributesNew}
\description{Creates a \code{\link{GtkTextAttributes}}, which describes
a set of properties on some text.}
\usage{gtkTextAttributesNew()}
\value{[\code{\link{GtkTextAttributes}}] a new \code{\link{GtkTextAttributes}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
