\alias{gtkTextBufferCopyClipboard}
\name{gtkTextBufferCopyClipboard}
\title{gtkTextBufferCopyClipboard}
\description{Copies the currently-selected text to a clipboard.}
\usage{gtkTextBufferCopyClipboard(object, clipboard)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTextBuffer}}]  a \code{\link{GtkTextBuffer}} }
\item{\code{clipboard}}{[\code{\link{GtkClipboard}}]  the \code{\link{GtkClipboard}} object to copy to.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
