\alias{gtkTextBufferCopyClipboard}
\name{gtkTextBufferCopyClipboard}
\title{gtkTextBufferCopyClipboard}
\description{Copies the currently-selected text to a clipboard.}
\usage{gtkTextBufferCopyClipboard(object, clipboard)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextBuffer}}}
\item{\verb{clipboard}}{the \code{\link{GtkClipboard}} object to copy to}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
