\alias{gtkTextBufferCutClipboard}
\name{gtkTextBufferCutClipboard}
\title{gtkTextBufferCutClipboard}
\description{Copies the currently-selected text to a clipboard, then deletes
said text if it's editable.}
\usage{gtkTextBufferCutClipboard(object, clipboard, default.editable)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTextBuffer}}]  a \code{\link{GtkTextBuffer}}}
\item{\code{clipboard}}{[\code{\link{GtkClipboard}}]  the \code{\link{GtkClipboard}} object to cut to}
\item{\code{default.editable}}{[logical]  default editability of the buffer}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
