\alias{gtkTextBufferCutClipboard}
\name{gtkTextBufferCutClipboard}
\title{gtkTextBufferCutClipboard}
\description{Copies the currently-selected text to a clipboard, then deletes
said text if it's editable.}
\usage{gtkTextBufferCutClipboard(object, clipboard, default.editable)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextBuffer}}}
\item{\verb{clipboard}}{the \code{\link{GtkClipboard}} object to cut to}
\item{\verb{default.editable}}{default editability of the buffer}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
