\alias{gtkTextBufferDeleteMarkByName}
\name{gtkTextBufferDeleteMarkByName}
\title{gtkTextBufferDeleteMarkByName}
\description{Deletes the mark named \code{name}; the mark must exist. See
\code{\link{gtkTextBufferDeleteMark}} for details.}
\usage{gtkTextBufferDeleteMarkByName(object, name)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextBuffer}}}
\item{\verb{name}}{name of a mark in \code{buffer}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
