\alias{gtkTextBufferGetHasSelection}
\name{gtkTextBufferGetHasSelection}
\title{gtkTextBufferGetHasSelection}
\description{Indicates whether the buffer has some text currently selected.}
\usage{gtkTextBufferGetHasSelection(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTextBuffer}}]  a \verb{\link{GtkTextBuffer}} }}
\details{  Since  2.10}
\value{[logical] \code{TRUE} if the there is text selected}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
