\alias{gtkTextBufferGetIterAtChildAnchor}
\name{gtkTextBufferGetIterAtChildAnchor}
\title{gtkTextBufferGetIterAtChildAnchor}
\description{Obtains the location of \code{anchor} within \code{buffer}.}
\usage{gtkTextBufferGetIterAtChildAnchor(object, anchor)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextBuffer}}}
\item{\verb{anchor}}{a child anchor that appears in \code{buffer}}
}
\value{
A list containing the following elements:
\item{\verb{iter}}{an iterator to be initialized}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
