\alias{gtkTextBufferGetIterAtLine}
\name{gtkTextBufferGetIterAtLine}
\title{gtkTextBufferGetIterAtLine}
\description{Initializes \code{iter} to the start of the given line.}
\usage{gtkTextBufferGetIterAtLine(object, line.number)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextBuffer}}}
\item{\verb{line.number}}{line number counting from 0}
}
\value{
A list containing the following elements:
\item{\verb{iter}}{iterator to initialize}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
