\alias{gtkTextBufferGetIterAtMark}
\name{gtkTextBufferGetIterAtMark}
\title{gtkTextBufferGetIterAtMark}
\description{Initializes \code{iter} with the current position of \code{mark}.}
\usage{gtkTextBufferGetIterAtMark(object, mark)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTextBuffer}}]  a \verb{\link{GtkTextBuffer}}}
\item{\verb{mark}}{[\verb{\link{GtkTextMark}}]  a \verb{\link{GtkTextMark}} in \code{buffer}}
}
\value{
A list containing the following elements:
\item{\verb{iter}}{[\verb{\link{GtkTextIter}}]  iterator to initialize}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
