\alias{gtkTextBufferGetIterAtMark}
\name{gtkTextBufferGetIterAtMark}
\title{gtkTextBufferGetIterAtMark}
\description{Initializes \code{iter} with the current position of \code{mark}.}
\usage{gtkTextBufferGetIterAtMark(object, mark)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextBuffer}}}
\item{\verb{mark}}{a \code{\link{GtkTextMark}} in \code{buffer}}
}
\value{
A list containing the following elements:
\item{\verb{iter}}{iterator to initialize}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
