\alias{gtkTextBufferGetTagTable}
\name{gtkTextBufferGetTagTable}
\title{gtkTextBufferGetTagTable}
\description{Get the \verb{\link{GtkTextTagTable}} associated with this buffer.}
\usage{gtkTextBufferGetTagTable(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTextBuffer}}]  a \verb{\link{GtkTextBuffer}}}}
\value{[\verb{\link{GtkTextTagTable}}]  the buffer's tag table}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
