\alias{gtkTextBufferGetTagTable}
\name{gtkTextBufferGetTagTable}
\title{gtkTextBufferGetTagTable}
\description{Get the \code{\link{GtkTextTagTable}} associated with this buffer.}
\usage{gtkTextBufferGetTagTable(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextBuffer}}}}
\value{[\code{\link{GtkTextTagTable}}] the buffer's tag table. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
