\alias{gtkTextBufferInsertWithTagsByName}
\name{gtkTextBufferInsertWithTagsByName}
\title{gtkTextBufferInsertWithTagsByName}
\description{Same as \code{\link{gtkTextBufferInsertWithTags}}, but allows you
to pass in tag names instead of tag objects.}
\usage{gtkTextBufferInsertWithTagsByName(object, iter, text, ...)}
\arguments{\item{\code{...}}{[\code{\link{GtkTextIter}}]  position in \code{buffer}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
