\alias{gtkTextBufferInsertWithTagsByName}
\name{gtkTextBufferInsertWithTagsByName}
\title{gtkTextBufferInsertWithTagsByName}
\description{Same as \code{\link{gtkTextBufferInsertWithTags}}, but allows you
to pass in tag names instead of tag objects.}
\usage{gtkTextBufferInsertWithTagsByName(object, iter, text, ...)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextBuffer}}}
\item{\verb{iter}}{position in \code{buffer}}
\item{\verb{text}}{UTF-8 text}
\item{\verb{...}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
