\alias{gtkTextBufferMoveMark}
\name{gtkTextBufferMoveMark}
\title{gtkTextBufferMoveMark}
\description{Moves \code{mark} to the new location \code{where}. Emits the "mark-set" signal
as notification of the move.}
\usage{gtkTextBufferMoveMark(object, mark, where)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextBuffer}}}
\item{\verb{mark}}{a \code{\link{GtkTextMark}}}
\item{\verb{where}}{new location for \code{mark} in \code{buffer}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
