\alias{gtkTextBufferMoveMarkByName}
\name{gtkTextBufferMoveMarkByName}
\title{gtkTextBufferMoveMarkByName}
\description{Moves the mark named \code{name} (which must exist) to location \code{where}.
See \code{\link{gtkTextBufferMoveMark}} for details.}
\usage{gtkTextBufferMoveMarkByName(object, name, where)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTextBuffer}}]  a \verb{\link{GtkTextBuffer}}}
\item{\verb{name}}{[character]  name of a mark}
\item{\verb{where}}{[\verb{\link{GtkTextIter}}]  new location for mark}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
