\alias{gtkTextBufferMoveMarkByName}
\name{gtkTextBufferMoveMarkByName}
\title{gtkTextBufferMoveMarkByName}
\description{Moves the mark named \code{name} (which must exist) to location \code{where}.
See \code{\link{gtkTextBufferMoveMark}} for details.}
\usage{gtkTextBufferMoveMarkByName(object, name, where)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextBuffer}}}
\item{\verb{name}}{name of a mark}
\item{\verb{where}}{new location for mark}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
