\alias{gtkTextBufferNew}
\name{gtkTextBufferNew}
\title{gtkTextBufferNew}
\description{Creates a new text buffer.}
\usage{gtkTextBufferNew(table = NULL)}
\arguments{\item{\code{table}}{[\code{\link{GtkTextTagTable}}]  a tag table, or \code{NULL} to create a new one}}
\value{[\code{\link{GtkTextBuffer}}]  a new text buffer}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
