\alias{gtkTextBufferNew}
\name{gtkTextBufferNew}
\title{gtkTextBufferNew}
\description{Creates a new text buffer.}
\usage{gtkTextBufferNew(table = NULL)}
\arguments{\item{\verb{table}}{a tag table, or \code{NULL} to create a new one. \emph{[  \acronym{allow-none}  ]}}}
\value{[\code{\link{GtkTextBuffer}}] a new text buffer}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
