\alias{gtkTextBufferRemoveSelectionClipboard}
\name{gtkTextBufferRemoveSelectionClipboard}
\title{gtkTextBufferRemoveSelectionClipboard}
\description{Removes a \code{\link{GtkClipboard}} added with 
\code{\link{gtkTextBufferAddSelectionClipboard}}.}
\usage{gtkTextBufferRemoveSelectionClipboard(object, clipboard)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextBuffer}}}
\item{\verb{clipboard}}{a \code{\link{GtkClipboard}} added to \code{buffer} by
\code{\link{gtkTextBufferAddSelectionClipboard}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
