\alias{gtkTextBufferUnregisterSerializeFormat}
\name{gtkTextBufferUnregisterSerializeFormat}
\title{gtkTextBufferUnregisterSerializeFormat}
\description{This function unregisters a rich text format that was previously
registered using \code{\link{gtkTextBufferRegisterSerializeFormat}} or
\code{\link{gtkTextBufferRegisterSerializeTagset}}}
\usage{gtkTextBufferUnregisterSerializeFormat(object, format)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextBuffer}}}
\item{\verb{format}}{a \code{\link{GdkAtom}} representing a registered rich text format.}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
