\alias{gtkTextChildAnchorGetDeleted}
\name{gtkTextChildAnchorGetDeleted}
\title{gtkTextChildAnchorGetDeleted}
\description{Determines whether a child anchor has been deleted from
the buffer. Keep in mind that the child anchor will be
unreferenced when removed from the buffer,}
\usage{gtkTextChildAnchorGetDeleted(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextChildAnchor}}}}
\value{[logical] \code{TRUE} if the child anchor has been deleted from its buffer}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
