\alias{gtkTextChildAnchorGetWidgets}
\name{gtkTextChildAnchorGetWidgets}
\title{gtkTextChildAnchorGetWidgets}
\description{Gets a list of all widgets anchored at this child anchor.}
\usage{gtkTextChildAnchorGetWidgets(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTextChildAnchor}}]  a \verb{\link{GtkTextChildAnchor}}}}
\value{[list]  list of widgets anchored at \code{anchor}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
