\alias{gtkTextChildAnchorGetWidgets}
\name{gtkTextChildAnchorGetWidgets}
\title{gtkTextChildAnchorGetWidgets}
\description{Gets a list of all widgets anchored at this child anchor.}
\usage{gtkTextChildAnchorGetWidgets(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextChildAnchor}}}}
\value{[list] list of widgets anchored at \code{anchor}. \emph{[  \acronym{element-type}   GtkWidget][  \acronym{transfer container}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
