\alias{gtkTextIterBackwardCursorPosition}
\name{gtkTextIterBackwardCursorPosition}
\title{gtkTextIterBackwardCursorPosition}
\description{Like \code{\link{gtkTextIterForwardCursorPosition}}, but moves backward.}
\usage{gtkTextIterBackwardCursorPosition(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTextIter}}]  a \verb{\link{GtkTextIter}}}}
\value{[logical] \code{TRUE} if we moved}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
