\alias{gtkTextIterBackwardVisibleWordStarts}
\name{gtkTextIterBackwardVisibleWordStarts}
\title{gtkTextIterBackwardVisibleWordStarts}
\description{Calls \code{\link{gtkTextIterBackwardVisibleWordStart}} up to \code{count} times.}
\usage{gtkTextIterBackwardVisibleWordStarts(object, count)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextIter}}}
\item{\verb{count}}{number of times to move}
}
\details{Since 2.4}
\value{[logical] \code{TRUE} if \code{iter} moved and is not the end iterator}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
