\alias{gtkTextIterGetBuffer}
\name{gtkTextIterGetBuffer}
\title{gtkTextIterGetBuffer}
\description{Returns the \code{\link{GtkTextBuffer}} this iterator is associated with.}
\usage{gtkTextIterGetBuffer(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTextIter}}]  an iterator}}
\value{[\code{\link{GtkTextBuffer}}]  the buffer}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
