\alias{gtkTextIterGetBuffer}
\name{gtkTextIterGetBuffer}
\title{gtkTextIterGetBuffer}
\description{Returns the \verb{\link{GtkTextBuffer}} this iterator is associated with.}
\usage{gtkTextIterGetBuffer(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTextIter}}]  an iterator}}
\value{[\verb{\link{GtkTextBuffer}}]  the buffer}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
