\alias{gtkTextIterGetBuffer}
\name{gtkTextIterGetBuffer}
\title{gtkTextIterGetBuffer}
\description{Returns the \code{\link{GtkTextBuffer}} this iterator is associated with.}
\usage{gtkTextIterGetBuffer(object)}
\arguments{\item{\verb{object}}{an iterator}}
\value{[\code{\link{GtkTextBuffer}}] the buffer. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
