\alias{gtkTextIterGetChar}
\name{gtkTextIterGetChar}
\title{gtkTextIterGetChar}
\description{returns 0.}
\usage{gtkTextIterGetChar(object)}
\arguments{\item{\verb{object}}{an iterator}}
\value{[numeric] a Unicode character, or 0 if \code{iter} is not dereferenceable}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
