\alias{gtkTextIterGetCharsInLine}
\name{gtkTextIterGetCharsInLine}
\title{gtkTextIterGetCharsInLine}
\description{Returns the number of characters in the line containing \code{iter},
including the paragraph delimiters.}
\usage{gtkTextIterGetCharsInLine(object)}
\arguments{\item{\verb{object}}{an iterator}}
\value{[integer] number of characters in the line}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
