\alias{gtkTextIterHasTag}
\name{gtkTextIterHasTag}
\title{gtkTextIterHasTag}
\description{Returns \code{TRUE} if \code{iter} is within a range tagged with \code{tag}.}
\usage{gtkTextIterHasTag(object, tag)}
\arguments{
\item{\verb{object}}{an iterator}
\item{\verb{tag}}{a \code{\link{GtkTextTag}}}
}
\value{[logical] whether \code{iter} is tagged with \code{tag}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
