\alias{gtkTextMarkGetBuffer}
\name{gtkTextMarkGetBuffer}
\title{gtkTextMarkGetBuffer}
\description{Gets the buffer this mark is located inside,
or \code{NULL} if the mark is deleted.}
\usage{gtkTextMarkGetBuffer(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTextMark}}]  a \verb{\link{GtkTextMark}}}}
\value{[\verb{\link{GtkTextBuffer}}]  the mark's \verb{\link{GtkTextBuffer}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
