\alias{gtkTextMarkGetBuffer}
\name{gtkTextMarkGetBuffer}
\title{gtkTextMarkGetBuffer}
\description{Gets the buffer this mark is located inside,
or \code{NULL} if the mark is deleted.}
\usage{gtkTextMarkGetBuffer(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextMark}}}}
\value{[\code{\link{GtkTextBuffer}}] the mark's \code{\link{GtkTextBuffer}}. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
