\alias{gtkTextMarkGetDeleted}
\name{gtkTextMarkGetDeleted}
\title{gtkTextMarkGetDeleted}
\description{Returns \code{TRUE} if the mark has been removed from its buffer
with \code{\link{gtkTextBufferDeleteMark}}. See \code{\link{gtkTextBufferAddMark}}
for a way to add it to a buffer again.}
\usage{gtkTextMarkGetDeleted(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextMark}}}}
\value{[logical] whether the mark is deleted}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
