\alias{gtkTextMarkGetName}
\name{gtkTextMarkGetName}
\title{gtkTextMarkGetName}
\description{Returns the mark name; returns NULL for anonymous marks.}
\usage{gtkTextMarkGetName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextMark}}}}
\value{[character] mark name}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
