\alias{gtkTextMarkGetVisible}
\name{gtkTextMarkGetVisible}
\title{gtkTextMarkGetVisible}
\description{Returns \code{TRUE} if the mark is visible (i.e. a cursor is displayed
for it).}
\usage{gtkTextMarkGetVisible(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextMark}}}}
\value{[logical] \code{TRUE} if visible}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
