\alias{gtkTextTagGetPriority}
\name{gtkTextTagGetPriority}
\title{gtkTextTagGetPriority}
\description{Get the tag priority.}
\usage{gtkTextTagGetPriority(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextTag}}}}
\value{[integer] The tag's priority.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
