\alias{gtkTextTagNew}
\name{gtkTextTagNew}
\title{gtkTextTagNew}
\description{Creates a \verb{\link{GtkTextTag}}. Configure the tag using object arguments,
i.e. using \code{\link{gObjectSet}}.}
\usage{gtkTextTagNew(name = NULL)}
\arguments{\item{\verb{name}}{[character]  tag name, or \code{NULL}}}
\value{[\verb{\link{GtkTextTag}}]  a new \verb{\link{GtkTextTag}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
