\alias{gtkTextTagNew}
\name{gtkTextTagNew}
\title{gtkTextTagNew}
\description{Creates a \code{\link{GtkTextTag}}. Configure the tag using object arguments,
i.e. using \code{\link{gObjectSet}}.}
\usage{gtkTextTagNew(name = NULL)}
\arguments{\item{\verb{name}}{tag name, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}}
\value{[\code{\link{GtkTextTag}}] a new \code{\link{GtkTextTag}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
