\alias{gtkTextTagTableLookup}
\name{gtkTextTagTableLookup}
\title{gtkTextTagTableLookup}
\description{Look up a named tag.}
\usage{gtkTextTagTableLookup(object, name)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTextTagTable}}]  a \code{\link{GtkTextTagTable}} }
\item{\code{name}}{[character]  name of a tag}
}
\value{[\code{\link{GtkTextTag}}]  The tag, or \code{NULL} if none by that name is in the table.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
