\alias{gtkTextTagTableLookup}
\name{gtkTextTagTableLookup}
\title{gtkTextTagTableLookup}
\description{Look up a named tag.}
\usage{gtkTextTagTableLookup(object, name)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextTagTable}}}
\item{\verb{name}}{name of a tag}
}
\value{[\code{\link{GtkTextTag}}] The tag, or \code{NULL} if none by that name is in the table. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
