\alias{gtkTextTagTableNew}
\name{gtkTextTagTableNew}
\title{gtkTextTagTableNew}
\description{Creates a new \verb{\link{GtkTextTagTable}}. The table contains no tags by
default.}
\usage{gtkTextTagTableNew()}
\value{[\verb{\link{GtkTextTagTable}}]  a new \verb{\link{GtkTextTagTable}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
