\alias{gtkTextViewAddChildAtAnchor}
\name{gtkTextViewAddChildAtAnchor}
\title{gtkTextViewAddChildAtAnchor}
\description{Adds a child widget in the text buffer, at the given \code{anchor}.}
\usage{gtkTextViewAddChildAtAnchor(object, child, anchor)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTextView}}]  a \code{\link{GtkTextView}}}
\item{\code{child}}{[\code{\link{GtkWidget}}]  a \code{\link{GtkWidget}}}
\item{\code{anchor}}{[\code{\link{GtkTextChildAnchor}}]  a \code{\link{GtkTextChildAnchor}} in the \code{\link{GtkTextBuffer}} for \code{text.view}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
