\alias{gtkTextViewAddChildAtAnchor}
\name{gtkTextViewAddChildAtAnchor}
\title{gtkTextViewAddChildAtAnchor}
\description{Adds a child widget in the text buffer, at the given \code{anchor}.}
\usage{gtkTextViewAddChildAtAnchor(object, child, anchor)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextView}}}
\item{\verb{child}}{a \code{\link{GtkWidget}}}
\item{\verb{anchor}}{a \code{\link{GtkTextChildAnchor}} in the \code{\link{GtkTextBuffer}} for \code{text.view}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
