\alias{gtkTextViewGetBuffer}
\name{gtkTextViewGetBuffer}
\title{gtkTextViewGetBuffer}
\description{Returns the \verb{\link{GtkTextBuffer}} being displayed by this text view.
The reference count on the buffer is not incremented; the caller
of this function won't own a new reference.}
\usage{gtkTextViewGetBuffer(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTextView}}]  a \verb{\link{GtkTextView}}}}
\value{[\verb{\link{GtkTextBuffer}}]  a \verb{\link{GtkTextBuffer}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
