\alias{gtkTextViewGetCursorVisible}
\name{gtkTextViewGetCursorVisible}
\title{gtkTextViewGetCursorVisible}
\description{Find out whether the cursor is being displayed.}
\usage{gtkTextViewGetCursorVisible(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextView}}}}
\value{[logical] whether the insertion mark is visible}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
