\alias{gtkTextViewGetLeftMargin}
\name{gtkTextViewGetLeftMargin}
\title{gtkTextViewGetLeftMargin}
\description{Gets the default left margin size of paragraphs in the \code{text.view}.
Tags in the buffer may override the default.}
\usage{gtkTextViewGetLeftMargin(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextView}}}}
\value{[integer] left margin in pixels}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
