\alias{gtkTextViewGetOverwrite}
\name{gtkTextViewGetOverwrite}
\title{gtkTextViewGetOverwrite}
\description{Returns whether the \verb{\link{GtkTextView}} is in overwrite mode or not.}
\usage{gtkTextViewGetOverwrite(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTextView}}]  a \verb{\link{GtkTextView}}}}
\details{  Since  2.4}
\value{[logical]  whether \code{text.view} is in overwrite mode or not.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
