\alias{gtkTextViewGetOverwrite}
\name{gtkTextViewGetOverwrite}
\title{gtkTextViewGetOverwrite}
\description{Returns whether the \code{\link{GtkTextView}} is in overwrite mode or not.}
\usage{gtkTextViewGetOverwrite(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextView}}}}
\details{Since 2.4}
\value{[logical] whether \code{text.view} is in overwrite mode or not.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
