\alias{gtkTextViewGetPixelsBelowLines}
\name{gtkTextViewGetPixelsBelowLines}
\title{gtkTextViewGetPixelsBelowLines}
\description{Gets the value set by \code{\link{gtkTextViewSetPixelsBelowLines}}.}
\usage{gtkTextViewGetPixelsBelowLines(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextView}}}}
\value{[integer] default number of blank pixels below paragraphs}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
