\alias{gtkTextViewGetPixelsInsideWrap}
\name{gtkTextViewGetPixelsInsideWrap}
\title{gtkTextViewGetPixelsInsideWrap}
\description{Gets the value set by \code{\link{gtkTextViewSetPixelsInsideWrap}}.}
\usage{gtkTextViewGetPixelsInsideWrap(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextView}}}}
\value{[integer] default number of pixels of blank space between wrapped lines}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
