\alias{gtkTextViewMoveChild}
\name{gtkTextViewMoveChild}
\title{gtkTextViewMoveChild}
\description{Updates the position of a child, as for \code{\link{gtkTextViewAddChildInWindow}}.}
\usage{gtkTextViewMoveChild(object, child, xpos, ypos)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextView}}}
\item{\verb{child}}{child widget already added to the text view}
\item{\verb{xpos}}{new X position in window coordinates}
\item{\verb{ypos}}{new Y position in window coordinates}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
