\alias{gtkTextViewPlaceCursorOnscreen}
\name{gtkTextViewPlaceCursorOnscreen}
\title{gtkTextViewPlaceCursorOnscreen}
\description{Moves the cursor to the currently visible region of the
buffer, it it isn't there already.}
\usage{gtkTextViewPlaceCursorOnscreen(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTextView}}}}
\value{[logical] \code{TRUE} if the cursor had to be moved.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
