\alias{gtkTextViewScrollMarkOnscreen}
\name{gtkTextViewScrollMarkOnscreen}
\title{gtkTextViewScrollMarkOnscreen}
\description{Scrolls \code{text.view} the minimum distance such that \code{mark} is contained
within the visible area of the widget.}
\usage{gtkTextViewScrollMarkOnscreen(object, mark)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextView}}}
\item{\verb{mark}}{a mark in the buffer for \code{text.view}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
