\alias{gtkTextViewSetCursorVisible}
\name{gtkTextViewSetCursorVisible}
\title{gtkTextViewSetCursorVisible}
\description{Toggles whether the insertion point is displayed. A buffer with no editable
text probably shouldn't have a visible cursor, so you may want to turn
the cursor off.}
\usage{gtkTextViewSetCursorVisible(object, setting)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTextView}}]  a \verb{\link{GtkTextView}}}
\item{\verb{setting}}{[logical]  whether to show the insertion cursor}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
