\alias{gtkTextViewSetEditable}
\name{gtkTextViewSetEditable}
\title{gtkTextViewSetEditable}
\description{Sets the default editability of the \code{\link{GtkTextView}}. You can override
this default setting with tags in the buffer, using the "editable"
attribute of tags.}
\usage{gtkTextViewSetEditable(object, setting)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTextView}}]  a \code{\link{GtkTextView}}}
\item{\code{setting}}{[logical]  whether it's editable}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
