\alias{gtkTextViewSetJustification}
\name{gtkTextViewSetJustification}
\title{gtkTextViewSetJustification}
\description{Sets the default justification of text in \code{text.view}.
Tags in the view's buffer may override the default.}
\usage{gtkTextViewSetJustification(object, justification)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTextView}}]  a \verb{\link{GtkTextView}}}
\item{\verb{justification}}{[\code{\link{GtkJustification}}]  justification}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
