\alias{gtkTextViewSetJustification}
\name{gtkTextViewSetJustification}
\title{gtkTextViewSetJustification}
\description{Sets the default justification of text in \code{text.view}.
Tags in the view's buffer may override the default.}
\usage{gtkTextViewSetJustification(object, justification)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextView}}}
\item{\verb{justification}}{justification}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
