\alias{gtkTextViewSetOverwrite}
\name{gtkTextViewSetOverwrite}
\title{gtkTextViewSetOverwrite}
\description{Changes the \verb{\link{GtkTextView}} overwrite mode.}
\usage{gtkTextViewSetOverwrite(object, overwrite)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTextView}}]  a \verb{\link{GtkTextView}}}
\item{\verb{overwrite}}{[logical] \code{TRUE} to turn on overwrite mode, \code{FALSE} to turn it off}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
