\alias{gtkTextViewSetOverwrite}
\name{gtkTextViewSetOverwrite}
\title{gtkTextViewSetOverwrite}
\description{Changes the \code{\link{GtkTextView}} overwrite mode.}
\usage{gtkTextViewSetOverwrite(object, overwrite)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextView}}}
\item{\verb{overwrite}}{\code{TRUE} to turn on overwrite mode, \code{FALSE} to turn it off}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
