\alias{gtkTextViewSetRightMargin}
\name{gtkTextViewSetRightMargin}
\title{gtkTextViewSetRightMargin}
\description{Sets the default right margin for text in the text view.
Tags in the buffer may override the default.}
\usage{gtkTextViewSetRightMargin(object, right.margin)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTextView}}]  a \code{\link{GtkTextView}}}
\item{\code{right.margin}}{[integer]  right margin in pixels}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
