\alias{gtkTextViewSetTabs}
\name{gtkTextViewSetTabs}
\title{gtkTextViewSetTabs}
\description{Sets the default tab stops for paragraphs in \code{text.view}.
Tags in the buffer may override the default.}
\usage{gtkTextViewSetTabs(object, tabs)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTextView}}]  a \verb{\link{GtkTextView}}}
\item{\verb{tabs}}{[\verb{\link{PangoTabArray}}]  tabs as a \verb{\link{PangoTabArray}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
