\alias{gtkTextViewSetTabs}
\name{gtkTextViewSetTabs}
\title{gtkTextViewSetTabs}
\description{Sets the default tab stops for paragraphs in \code{text.view}.
Tags in the buffer may override the default.}
\usage{gtkTextViewSetTabs(object, tabs)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextView}}}
\item{\verb{tabs}}{tabs as a \code{\link{PangoTabArray}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
