\alias{gtkTextViewSetWrapMode}
\name{gtkTextViewSetWrapMode}
\title{gtkTextViewSetWrapMode}
\description{Sets the line wrapping for the view.}
\usage{gtkTextViewSetWrapMode(object, wrap.mode)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTextView}}}
\item{\verb{wrap.mode}}{a \code{\link{GtkWrapMode}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
