\alias{gtkTipsQueryStartQuery}
\name{gtkTipsQueryStartQuery}
\title{gtkTipsQueryStartQuery}
\description{
Starts a query.
The \code{\link{GtkTipsQuery}} widget will take control of the mouse and as the mouse
moves it will display the tooltip of the widget beneath the mouse.
\strong{WARNING: \code{gtk_tips_query_start_query} is deprecated and should not be used in newly-written code.}
}
\usage{gtkTipsQueryStartQuery(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTipsQuery}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
