\alias{gtkToggleActionGetDrawAsRadio}
\name{gtkToggleActionGetDrawAsRadio}
\title{gtkToggleActionGetDrawAsRadio}
\description{Returns whether the action should have proxies like a radio action.}
\usage{gtkToggleActionGetDrawAsRadio(object)}
\arguments{\item{\verb{object}}{the action object}}
\details{Since 2.4}
\value{[logical] whether the action should have proxies like a radio action.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
