\alias{gtkToggleActionSetDrawAsRadio}
\name{gtkToggleActionSetDrawAsRadio}
\title{gtkToggleActionSetDrawAsRadio}
\description{Sets whether the action should have proxies like a radio action.}
\usage{gtkToggleActionSetDrawAsRadio(object, draw.as.radio)}
\arguments{
\item{\verb{object}}{the action object}
\item{\verb{draw.as.radio}}{whether the action should have proxies like a radio
action}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
