\alias{gtkToggleButtonGetActive}
\name{gtkToggleButtonGetActive}
\title{gtkToggleButtonGetActive}
\description{Queries a \code{\link{GtkToggleButton}} and returns its current state. Returns \code{TRUE} if
the toggle button is pressed in and \code{FALSE} if it is raised.}
\usage{gtkToggleButtonGetActive(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToggleButton}}.}}
\value{[logical] a \verb{logical} value.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
