\alias{gtkToggleButtonGetInconsistent}
\name{gtkToggleButtonGetInconsistent}
\title{gtkToggleButtonGetInconsistent}
\description{Gets the value set by \code{\link{gtkToggleButtonSetInconsistent}}.}
\usage{gtkToggleButtonGetInconsistent(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToggleButton}}}}
\value{[logical] \code{TRUE} if the button is displayed as inconsistent, \code{FALSE} otherwise}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
