\alias{gtkToggleButtonSetActive}
\name{gtkToggleButtonSetActive}
\title{gtkToggleButtonSetActive}
\description{Sets the status of the toggle button. Set to \code{TRUE} if you want the
GtkToggleButton to be 'pressed in', and \code{FALSE} to raise it.
This action causes the toggled signal to be emitted.}
\usage{gtkToggleButtonSetActive(object, is.active)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToggleButton}}.}
\item{\verb{is.active}}{\code{TRUE} or \code{FALSE}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
