\alias{gtkToggleToolButtonNew}
\name{gtkToggleToolButtonNew}
\title{gtkToggleToolButtonNew}
\description{Returns a new \code{\link{GtkToggleToolButton}}}
\usage{gtkToggleToolButtonNew(show = TRUE)}
\details{Since 2.4}
\value{[\code{\link{GtkToolItem}}] a newly created \code{\link{GtkToggleToolButton}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
