\alias{gtkToolbarAppendSpace}
\name{gtkToolbarAppendSpace}
\title{gtkToolbarAppendSpace}
\description{
Adds a new space to the end of the toolbar.
\strong{WARNING: \code{gtk_toolbar_append_space} is deprecated and should not be used in newly-written code.}
}
\usage{gtkToolbarAppendSpace(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkToolbar}}]  a \verb{\link{GtkToolbar}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
