\alias{gtkToolbarGetIconSize}
\name{gtkToolbarGetIconSize}
\title{gtkToolbarGetIconSize}
\description{Retrieves the icon size for the toolbar. See \code{\link{gtkToolbarSetIconSize}}.}
\usage{gtkToolbarGetIconSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolbar}}}}
\value{[\code{\link{GtkIconSize}}] the current icon size for the icons on
the toolbar. \emph{[  \acronym{type}   int]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
