\alias{gtkToolbarGetOrientation}
\name{gtkToolbarGetOrientation}
\title{gtkToolbarGetOrientation}
\description{Retrieves the current orientation of the toolbar. See
\code{\link{gtkToolbarSetOrientation}}.}
\usage{gtkToolbarGetOrientation(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkToolbar}}]  a \code{\link{GtkToolbar}}}}
\value{[\code{\link{GtkOrientation}}]  the orientation}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
