\alias{gtkToolbarGetOrientation}
\name{gtkToolbarGetOrientation}
\title{gtkToolbarGetOrientation}
\description{
Retrieves the current orientation of the toolbar. See
\code{\link{gtkToolbarSetOrientation}}.
\strong{WARNING: \code{gtk_toolbar_get_orientation} has been deprecated since version 2.16 and should not be used in newly-written code. Use \code{\link{gtkOrientableGetOrientation}} instead.}
}
\usage{gtkToolbarGetOrientation(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolbar}}}}
\value{[\code{\link{GtkOrientation}}] the orientation}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
