\alias{gtkToolbarGetReliefStyle}
\name{gtkToolbarGetReliefStyle}
\title{gtkToolbarGetReliefStyle}
\description{Returns the relief style of buttons on \code{toolbar}. See
\code{\link{gtkButtonSetRelief}}.}
\usage{gtkToolbarGetReliefStyle(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolbar}}}}
\details{Since 2.4}
\value{[\code{\link{GtkReliefStyle}}] The relief style of buttons on \code{toolbar}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
