\alias{gtkToolbarGetTooltips}
\name{gtkToolbarGetTooltips}
\title{gtkToolbarGetTooltips}
\description{
Retrieves whether tooltips are enabled. See
\code{\link{gtkToolbarSetTooltips}}.
\strong{WARNING: \code{gtk_toolbar_get_tooltips} has been deprecated since version 2.14 and should not be used in newly-written code. The toolkit-wide \verb{"gtk-enable-tooltips"} property
is now used instead.}
}
\usage{gtkToolbarGetTooltips(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolbar}}}}
\value{[logical] \code{TRUE} if tooltips are enabled}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
