\alias{gtkToolbarInsert}
\name{gtkToolbarInsert}
\title{gtkToolbarInsert}
\description{Insert a \code{\link{GtkToolItem}} into the toolbar at position \code{pos}. If \code{pos} is
0 the item is prepended to the start of the toolbar. If \code{pos} is
negative, the item is appended to the end of the toolbar.}
\usage{gtkToolbarInsert(object, item, pos)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolbar}}}
\item{\verb{item}}{a \code{\link{GtkToolItem}}}
\item{\verb{pos}}{the position of the new item}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
