\alias{gtkToolbarInsertSpace}
\name{gtkToolbarInsertSpace}
\title{gtkToolbarInsertSpace}
\description{
Inserts a new space in the toolbar at the specified position.
\strong{WARNING: \code{gtk_toolbar_insert_space} is deprecated and should not be used in newly-written code.}
}
\usage{gtkToolbarInsertSpace(object, position)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkToolbar}}]  a \verb{\link{GtkToolbar}}}
\item{\verb{position}}{[integer]  the number of widgets after which a space should be inserted.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
