\alias{gtkToolbarInsertSpace}
\name{gtkToolbarInsertSpace}
\title{gtkToolbarInsertSpace}
\description{
Inserts a new space in the toolbar at the specified position.
\strong{WARNING: \code{gtk_toolbar_insert_space} has been deprecated since version 2.4 and should not be used in newly-written code. Use \code{\link{gtkToolbarInsert}} instead.}
}
\usage{gtkToolbarInsertSpace(object, position)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolbar}}}
\item{\verb{position}}{the number of widgets after which a space should be inserted.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
