\alias{gtkToolbarPrependSpace}
\name{gtkToolbarPrependSpace}
\title{gtkToolbarPrependSpace}
\description{
Adds a new space to the beginning of the toolbar.
\strong{WARNING: \code{gtk_toolbar_prepend_space} is deprecated and should not be used in newly-written code.  
Adds a new space to the beginning of the toolbar.  }
}
\usage{gtkToolbarPrependSpace(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkToolbar}}]  a \code{\link{GtkToolbar}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
