\alias{gtkToolbarPrependSpace}
\name{gtkToolbarPrependSpace}
\title{gtkToolbarPrependSpace}
\description{
Adds a new space to the beginning of the toolbar.
\strong{WARNING: \code{gtk_toolbar_prepend_space} has been deprecated since version 2.4 and should not be used in newly-written code. Use \code{\link{gtkToolbarInsert}} instead.}
}
\usage{gtkToolbarPrependSpace(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolbar}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
