\alias{gtkToolbarRemoveSpace}
\name{gtkToolbarRemoveSpace}
\title{gtkToolbarRemoveSpace}
\description{
Removes a space from the specified position.
\strong{WARNING: \code{gtk_toolbar_remove_space} is deprecated and should not be used in newly-written code.  
Removes a space from the specified position.  }
}
\usage{gtkToolbarRemoveSpace(object, position)}
\arguments{
\item{\code{object}}{[\code{\link{GtkToolbar}}]  a \code{\link{GtkToolbar}}.}
\item{\code{position}}{[integer]  the index of the space to remove.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
