\alias{gtkToolbarRemoveSpace}
\name{gtkToolbarRemoveSpace}
\title{gtkToolbarRemoveSpace}
\description{
Removes a space from the specified position.
\strong{WARNING: \code{gtk_toolbar_remove_space} has been deprecated since version 2.4 and should not be used in newly-written code. Use \code{\link{gtkToolbarInsert}} instead.}
}
\usage{gtkToolbarRemoveSpace(object, position)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolbar}}.}
\item{\verb{position}}{the index of the space to remove.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
