\alias{gtkToolbarSetOrientation}
\name{gtkToolbarSetOrientation}
\title{gtkToolbarSetOrientation}
\description{
Sets whether a toolbar should appear horizontally or vertically.
\strong{WARNING: \code{gtk_toolbar_set_orientation} has been deprecated since version 2.16 and should not be used in newly-written code. Use \code{\link{gtkOrientableSetOrientation}} instead.}
}
\usage{gtkToolbarSetOrientation(object, orientation)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolbar}}.}
\item{\verb{orientation}}{a new \code{\link{GtkOrientation}}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
