\alias{gtkToolbarSetShowArrow}
\name{gtkToolbarSetShowArrow}
\title{gtkToolbarSetShowArrow}
\description{Sets whether to show an overflow menu when
\code{toolbar} doesn't have room for all items on it. If \code{TRUE},
items that there are not room are available through an
overflow menu.}
\usage{gtkToolbarSetShowArrow(object, show.arrow)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolbar}}}
\item{\verb{show.arrow}}{Whether to show an overflow menu}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
