\alias{gtkToolbarSetStyle}
\name{gtkToolbarSetStyle}
\title{gtkToolbarSetStyle}
\description{Alters the view of \code{toolbar} to display either icons only, text only, or both.}
\usage{gtkToolbarSetStyle(object, style)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolbar}}.}
\item{\verb{style}}{the new style for \code{toolbar}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
