\alias{gtkToolbarSetTooltips}
\name{gtkToolbarSetTooltips}
\title{gtkToolbarSetTooltips}
\description{Sets if the tooltips of a toolbar should be active or not.}
\usage{gtkToolbarSetTooltips(object, enable)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkToolbar}}]  a \verb{\link{GtkToolbar}}.}
\item{\verb{enable}}{[logical]  set to \code{FALSE} to disable the tooltips, or \code{TRUE} to enable them.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
