\alias{gtkToolbarUnsetIconSize}
\name{gtkToolbarUnsetIconSize}
\title{gtkToolbarUnsetIconSize}
\description{
Unsets toolbar icon size set with \code{\link{gtkToolbarSetIconSize}}, so that
user preferences will be used to determine the icon size.
\strong{WARNING: \code{gtk_toolbar_unset_icon_size} is deprecated and should not be used in newly-written code.  
Unsets toolbar icon size set with \code{\link{gtkToolbarSetIconSize}}, so that
user preferences will be used to determine the icon size.  }
}
\usage{gtkToolbarUnsetIconSize(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkToolbar}}]  a \code{\link{GtkToolbar}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
