\alias{gtkToolbarUnsetIconSize}
\name{gtkToolbarUnsetIconSize}
\title{gtkToolbarUnsetIconSize}
\description{Unsets toolbar icon size set with \code{\link{gtkToolbarSetIconSize}}, so that
user preferences will be used to determine the icon size.}
\usage{gtkToolbarUnsetIconSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolbar}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
