\alias{gtkToolbarUnsetStyle}
\name{gtkToolbarUnsetStyle}
\title{gtkToolbarUnsetStyle}
\description{Unsets a toolbar style set with \code{\link{gtkToolbarSetStyle}}, so that
user preferences will be used to determine the toolbar style.}
\usage{gtkToolbarUnsetStyle(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkToolbar}}]  a \code{\link{GtkToolbar}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
