\alias{gtkToolButtonGetIconWidget}
\name{gtkToolButtonGetIconWidget}
\title{gtkToolButtonGetIconWidget}
\description{Return the widget used as icon widget on \code{button}. See
\code{\link{gtkToolButtonSetIconWidget}}.}
\usage{gtkToolButtonGetIconWidget(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolButton}}}}
\details{Since 2.4}
\value{[\code{\link{GtkWidget}}] The widget used as icon on \code{button}, or \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
