\alias{gtkToolButtonNew}
\name{gtkToolButtonNew}
\title{gtkToolButtonNew}
\description{Creates a new \code{\link{GtkToolButton}} using \code{icon.widget} as icon and \code{label} as
label.}
\usage{gtkToolButtonNew(icon.widget = NULL, label = NULL, show = TRUE)}
\arguments{
\item{\verb{icon.widget}}{a \code{\link{GtkMisc}} widget that will be used as icon widget, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
\item{\verb{label}}{a string that will be used as label, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.4}
\value{[\code{\link{GtkToolItem}}] A new \code{\link{GtkToolButton}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
