\alias{gtkToolItemGetExpand}
\name{gtkToolItemGetExpand}
\title{gtkToolItemGetExpand}
\description{Returns whether \code{tool.item} is allocated extra space.
See \code{\link{gtkToolItemSetExpand}}.}
\usage{gtkToolItemGetExpand(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolItem}}}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if \code{tool.item} is allocated extra space.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
