\alias{gtkToolItemGetIconSize}
\name{gtkToolItemGetIconSize}
\title{gtkToolItemGetIconSize}
\description{Returns the icon size used for \code{tool.item}. Custom subclasses of
\code{\link{GtkToolItem}} should call this function to find out what size icons
they should use.}
\usage{gtkToolItemGetIconSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolItem}}}}
\details{Since 2.4}
\value{[\code{\link{GtkIconSize}}] a \code{\link{GtkIconSize}} indicating the icon size
used for \code{tool.item}. \emph{[  \acronym{type}   int]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
