\alias{gtkToolItemGetUseDragWindow}
\name{gtkToolItemGetUseDragWindow}
\title{gtkToolItemGetUseDragWindow}
\description{Returns whether \code{tool.item} has a drag window. See
\code{\link{gtkToolItemSetUseDragWindow}}.}
\usage{gtkToolItemGetUseDragWindow(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolItem}}}}
\details{Since 2.4}
\value{[logical] \code{TRUE} if \code{tool.item} uses a drag window.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
