\alias{gtkToolItemNew}
\name{gtkToolItemNew}
\title{gtkToolItemNew}
\description{Creates a new \code{\link{GtkToolItem}}}
\usage{gtkToolItemNew(show = TRUE)}
\details{Since 2.4}
\value{[\code{\link{GtkToolItem}}] the new \code{\link{GtkToolItem}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
