\alias{gtkToolItemRetrieveProxyMenuItem}
\name{gtkToolItemRetrieveProxyMenuItem}
\title{gtkToolItemRetrieveProxyMenuItem}
\description{Returns the \code{\link{GtkMenuItem}} that was last set by
\code{\link{gtkToolItemSetProxyMenuItem}}, ie. the \code{\link{GtkMenuItem}}
that is going to appear in the overflow menu.}
\usage{gtkToolItemRetrieveProxyMenuItem(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkToolItem}}}}
\details{Since 2.4}
\value{[\code{\link{GtkWidget}}] The \code{\link{GtkMenuItem}} that is going to appear in the
overflow menu for \code{tool.item}. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
