\alias{gtkToolItemSetHomogeneous}
\name{gtkToolItemSetHomogeneous}
\title{gtkToolItemSetHomogeneous}
\description{Sets whether \code{tool.item} is to be allocated the same size as other
homogeneous items. The effect is that all homogeneous items will have
the same width as the widest of the items.}
\usage{gtkToolItemSetHomogeneous(object, homogeneous)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkToolItem}}]  a \code{\link{gtkToolItem}} }
\item{\verb{homogeneous}}{[logical]  whether \code{tool.item} is the same size as other homogeneous items}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
