\alias{gtkToolItemSetTooltipMarkup}
\name{gtkToolItemSetTooltipMarkup}
\title{gtkToolItemSetTooltipMarkup}
\description{Sets the markup text to be displayed as tooltip on the item.
See \code{\link{gtkWidgetSetTooltipMarkup}}.}
\usage{gtkToolItemSetTooltipMarkup(object, markup)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolItem}}}
\item{\verb{markup}}{markup text to be used as tooltip for \code{tool.item}}
}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
