\alias{gtkToolItemSetTooltipText}
\name{gtkToolItemSetTooltipText}
\title{gtkToolItemSetTooltipText}
\description{Sets the text to be displayed as tooltip on the item.
See \code{\link{gtkWidgetSetTooltipText}}.}
\usage{gtkToolItemSetTooltipText(object, text)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkToolItem}}}
\item{\verb{text}}{text to be used as tooltip for \code{tool.item}}
}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
