\alias{gtkToolItemSetVisibleHorizontal}
\name{gtkToolItemSetVisibleHorizontal}
\title{gtkToolItemSetVisibleHorizontal}
\description{Sets whether \code{tool.item} is visible when the toolbar is docked horizontally.}
\usage{gtkToolItemSetVisibleHorizontal(object, visible.horizontal)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkToolItem}}]  a \verb{\link{GtkToolItem}}}
\item{\verb{visible.horizontal}}{[logical]  Whether \code{tool.item} is visible when in horizontal mode}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
