\alias{gtkTooltipsDisable}
\name{gtkTooltipsDisable}
\title{gtkTooltipsDisable}
\description{Causes all tooltips in \code{tooltips} to become inactive. Any widgets that have tips associated with that group will no longer display their tips until they are enabled again with \code{\link{gtkTooltipsEnable}}.}
\usage{gtkTooltipsDisable(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTooltips}}] a \code{\link{GtkTooltips}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
