\alias{gtkTooltipsEnable}
\name{gtkTooltipsEnable}
\title{gtkTooltipsEnable}
\description{
Allows the user to see your tooltips as they navigate your application.
\strong{WARNING: \code{gtk_tooltips_enable} has been deprecated since version 2.12 and should not be used in newly-written code. }
}
\usage{gtkTooltipsEnable(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTooltips}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
