\alias{gtkTooltipSetIcon}
\name{gtkTooltipSetIcon}
\title{gtkTooltipSetIcon}
\description{Sets the icon of the tooltip (which is in front of the text) to be
\code{pixbuf}.  If \code{pixbuf} is \code{NULL}, the image will be hidden.}
\usage{gtkTooltipSetIcon(object, pixbuf)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTooltip}}]  a \verb{\link{GtkTooltip}}}
\item{\verb{pixbuf}}{[\verb{\link{GdkPixbuf}}]  a \verb{\link{GdkPixbuf}}, or \code{NULL}}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
