\alias{gtkTooltipSetIcon}
\name{gtkTooltipSetIcon}
\title{gtkTooltipSetIcon}
\description{Sets the icon of the tooltip (which is in front of the text) to be
\code{pixbuf}.  If \code{pixbuf} is \code{NULL}, the image will be hidden.}
\usage{gtkTooltipSetIcon(object, pixbuf)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTooltip}}}
\item{\verb{pixbuf}}{a \code{\link{GdkPixbuf}}, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
