\alias{gtkTooltipSetMarkup}
\name{gtkTooltipSetMarkup}
\title{gtkTooltipSetMarkup}
\description{Sets the text of the tooltip to be \code{markup}, which is marked up
with the Pango text markup language.
If \code{markup} is \code{NULL}, the label will be hidden.}
\usage{gtkTooltipSetMarkup(object, markup)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTooltip}}}
\item{\verb{markup}}{a markup string (see Pango markup format) or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
