\alias{gtkTooltipSetText}
\name{gtkTooltipSetText}
\title{gtkTooltipSetText}
\description{Sets the text of the tooltip to be \code{text}. If \code{text} is \code{NULL}, the label
will be hidden. See also \code{\link{gtkTooltipSetMarkup}}.}
\usage{gtkTooltipSetText(object, text)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTooltip}}}
\item{\verb{text}}{a text string or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
