\alias{gtkTooltipsNew}
\name{gtkTooltipsNew}
\title{gtkTooltipsNew}
\description{
Creates an empty group of tooltips. This function initialises a \code{\link{GtkTooltips}} structure. Without at least one such structure, you can not add tips to your application.
\strong{WARNING: \code{gtk_tooltips_new} has been deprecated since version 2.12 and should not be used in newly-written code. }
}
\usage{gtkTooltipsNew()}
\value{[\code{\link{GtkTooltips}}] a new \code{\link{GtkTooltips}} group for you to use.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
