\alias{gtkTooltipsSetDelay}
\name{gtkTooltipsSetDelay}
\title{gtkTooltipsSetDelay}
\description{
Sets the time between the user moving the mouse over a widget and the widget's tooltip appearing.
\strong{WARNING: \code{gtk_tooltips_set_delay} has been deprecated since version 2.12 and should not be used in newly-written code. }
}
\usage{gtkTooltipsSetDelay(object, delay)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTooltips}}.}
\item{\verb{delay}}{an integer value representing milliseconds.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
