\alias{gtkTreeModelFilterGetModel}
\name{gtkTreeModelFilterGetModel}
\title{gtkTreeModelFilterGetModel}
\description{Returns a pointer to the child model of \code{filter}.}
\usage{gtkTreeModelFilterGetModel(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeModelFilter}}.}}
\details{Since 2.4}
\value{[\code{\link{GtkTreeModel}}] A pointer to a \code{\link{GtkTreeModel}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
