\alias{gtkTreeModelFilterNew}
\name{gtkTreeModelFilterNew}
\title{gtkTreeModelFilterNew}
\description{Creates a new \code{\link{GtkTreeModel}}, with \code{child.model} as the child\_model
and \code{root} as the virtual root.}
\usage{gtkTreeModelFilterNew(child.model, root = NULL)}
\arguments{
\item{\code{child.model}}{[\code{\link{GtkTreeModel}}]  A \code{\link{GtkTreeModel}}.}
\item{\code{root}}{[\code{\link{GtkTreePath}}]  A \code{\link{GtkTreePath}} or \code{NULL}.}
}
\details{  Since  2.4}
\value{[\code{\link{GtkTreeModel}}]  A new \code{\link{GtkTreeModel}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
