\alias{gtkTreeModelFilterNew}
\name{gtkTreeModelFilterNew}
\title{gtkTreeModelFilterNew}
\description{Creates a new \code{\link{GtkTreeModel}}, with \code{child.model} as the child_model
and \code{root} as the virtual root.}
\usage{gtkTreeModelFilterNew(child.model, root = NULL)}
\arguments{
\item{\verb{child.model}}{A \code{\link{GtkTreeModel}}.}
\item{\verb{root}}{A \code{\link{GtkTreePath}} or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\details{Since 2.4}
\value{[\code{\link{GtkTreeModel}}] A new \code{\link{GtkTreeModel}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
