\alias{gtkTreeModelGetColumnType}
\name{gtkTreeModelGetColumnType}
\title{gtkTreeModelGetColumnType}
\description{Returns the type of the column.}
\usage{gtkTreeModelGetColumnType(object, index)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeModel}}]  A \code{\link{GtkTreeModel}}.}
\item{\code{index}}{[integer]  The column index.}
}
\value{[\code{\link{GType}}]  The type of the column.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
