\alias{gtkTreeModelGetColumnType}
\name{gtkTreeModelGetColumnType}
\title{gtkTreeModelGetColumnType}
\description{Returns the type of the column.}
\usage{gtkTreeModelGetColumnType(object, index)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeModel}}.}
\item{\verb{index}}{The column index.}
}
\value{[\code{\link{GType}}] The type of the column.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
