\alias{gtkTreeModelGetNColumns}
\name{gtkTreeModelGetNColumns}
\title{gtkTreeModelGetNColumns}
\description{Returns the number of columns supported by \code{tree.model}.}
\usage{gtkTreeModelGetNColumns(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeModel}}.}}
\value{[integer] The number of columns.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
