\alias{gtkTreeModelGetPath}
\name{gtkTreeModelGetPath}
\title{gtkTreeModelGetPath}
\description{Returns a newly-created \verb{\link{GtkTreePath}} referenced by \code{iter}.}
\usage{gtkTreeModelGetPath(object, iter)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeModel}}]  A \verb{\link{GtkTreeModel}}.}
\item{\verb{iter}}{[\verb{\link{GtkTreeIter}}]  The \verb{\link{GtkTreeIter}}.}
}
\value{[\verb{\link{GtkTreePath}}]  a newly-created \verb{\link{GtkTreePath}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
