\alias{gtkTreeModelGetPath}
\name{gtkTreeModelGetPath}
\title{gtkTreeModelGetPath}
\description{Returns a newly-created \code{\link{GtkTreePath}} referenced by \code{iter}.}
\usage{gtkTreeModelGetPath(object, iter)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeModel}}.}
\item{\verb{iter}}{The \code{\link{GtkTreeIter}}.}
}
\value{[\code{\link{GtkTreePath}}] a newly-created \code{\link{GtkTreePath}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
