\alias{gtkTreeModelIterHasChild}
\name{gtkTreeModelIterHasChild}
\title{gtkTreeModelIterHasChild}
\description{Returns \code{TRUE} if \code{iter} has children, \code{FALSE} otherwise.}
\usage{gtkTreeModelIterHasChild(object, iter)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeModel}}]  A \code{\link{GtkTreeModel}}.}
\item{\code{iter}}{[\code{\link{GtkTreeIter}}]  The \code{\link{GtkTreeIter}} to test for children.}
}
\value{[logical] \code{TRUE} if \code{iter} has children.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
