\alias{gtkTreeModelRowChanged}
\name{gtkTreeModelRowChanged}
\title{gtkTreeModelRowChanged}
\description{Emits the "row_changed" signal on \code{tree.model}.}
\usage{gtkTreeModelRowChanged(object, path, iter)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeModel}}]  A \verb{\link{GtkTreeModel}}}
\item{\verb{path}}{[\verb{\link{GtkTreePath}}]  A \verb{\link{GtkTreePath}} pointing to the changed row}
\item{\verb{iter}}{[\verb{\link{GtkTreeIter}}]  A valid \verb{\link{GtkTreeIter}} pointing to the changed row}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
