\alias{gtkTreeModelRowChanged}
\name{gtkTreeModelRowChanged}
\title{gtkTreeModelRowChanged}
\description{Emits the "row-changed" signal on \code{tree.model}.}
\usage{gtkTreeModelRowChanged(object, path, iter)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeModel}}}
\item{\verb{path}}{A \code{\link{GtkTreePath}} pointing to the changed row}
\item{\verb{iter}}{A valid \code{\link{GtkTreeIter}} pointing to the changed row}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
