\alias{gtkTreeModelRowInserted}
\name{gtkTreeModelRowInserted}
\title{gtkTreeModelRowInserted}
\description{Emits the "row\_inserted" signal on \code{tree.model}}
\usage{gtkTreeModelRowInserted(object, path, iter)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeModel}}]  A \code{\link{GtkTreeModel}}}
\item{\code{path}}{[\code{\link{GtkTreePath}}]  A \code{\link{GtkTreePath}} pointing to the inserted row}
\item{\code{iter}}{[\code{\link{GtkTreeIter}}]  A valid \code{\link{GtkTreeIter}} pointing to the inserted row}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
