\alias{gtkTreeModelRowInserted}
\name{gtkTreeModelRowInserted}
\title{gtkTreeModelRowInserted}
\description{Emits the "row-inserted" signal on \code{tree.model}}
\usage{gtkTreeModelRowInserted(object, path, iter)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeModel}}}
\item{\verb{path}}{A \code{\link{GtkTreePath}} pointing to the inserted row}
\item{\verb{iter}}{A valid \code{\link{GtkTreeIter}} pointing to the inserted row}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
