\alias{gtkTreeModelSortGetModel}
\name{gtkTreeModelSortGetModel}
\title{gtkTreeModelSortGetModel}
\description{Returns the model the \code{\link{GtkTreeModelSort}} is sorting.}
\usage{gtkTreeModelSortGetModel(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTreeModelSort}}}}
\value{[\code{\link{GtkTreeModel}}] the "child model" being sorted}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
