\alias{gtkTreeModelSortNewWithModel}
\name{gtkTreeModelSortNewWithModel}
\title{gtkTreeModelSortNewWithModel}
\description{Creates a new \code{\link{GtkTreeModel}}, with \code{child.model} as the child model.}
\usage{gtkTreeModelSortNewWithModel(child.model = NULL)}
\arguments{\item{\verb{child.model}}{A \code{\link{GtkTreeModel}}}}
\value{[\code{\link{GtkTreeModel}}] A new \code{\link{GtkTreeModel}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
