\alias{gtkTreePathCopy}
\name{gtkTreePathCopy}
\title{gtkTreePathCopy}
\description{Creates a new \verb{\link{GtkTreePath}} as a copy of \code{path}.}
\usage{gtkTreePathCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreePath}}]  A \verb{\link{GtkTreePath}}.}}
\value{[\verb{\link{GtkTreePath}}]  A new \verb{\link{GtkTreePath}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
