\alias{gtkTreePathDown}
\name{gtkTreePathDown}
\title{gtkTreePathDown}
\description{Moves \code{path} to point to the first child of the current path.}
\usage{gtkTreePathDown(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreePath}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
