\alias{gtkTreePathGetDepth}
\name{gtkTreePathGetDepth}
\title{gtkTreePathGetDepth}
\description{Returns the current depth of \code{path}.}
\usage{gtkTreePathGetDepth(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreePath}}]  A \code{\link{GtkTreePath}}.}}
\value{[integer]  The depth of \code{path}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
