\alias{gtkTreePathGetIndices}
\name{gtkTreePathGetIndices}
\title{gtkTreePathGetIndices}
\description{Returns the current indices of \code{path}.  This is a list of integers, each
representing a node in a tree.}
\usage{gtkTreePathGetIndices(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreePath}}]  A \code{\link{GtkTreePath}}.}}
\value{[integer]  The current indices, or \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
