\alias{gtkTreePathGetIndices}
\name{gtkTreePathGetIndices}
\title{gtkTreePathGetIndices}
\description{Returns the current indices of \code{path}.  This is a list of integers, each
representing a node in a tree.  This value should not be freed.}
\usage{gtkTreePathGetIndices(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreePath}}.}}
\value{[integer] The current indices, or \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
