\alias{gtkTreePathIsAncestor}
\name{gtkTreePathIsAncestor}
\title{gtkTreePathIsAncestor}
\description{Returns \code{TRUE} if \code{descendant} is a descendant of \code{path}.}
\usage{gtkTreePathIsAncestor(object, descendant)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreePath}}]  a \code{\link{GtkTreePath}}}
\item{\code{descendant}}{[\code{\link{GtkTreePath}}]  another \code{\link{GtkTreePath}}}
}
\value{[logical] \code{TRUE} if \code{descendant} is contained inside \code{path}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
