\alias{gtkTreePathNew}
\name{gtkTreePathNew}
\title{gtkTreePathNew}
\description{Creates a new \code{\link{GtkTreePath}}.  This structure refers to a row.}
\usage{gtkTreePathNew()}
\value{[\code{\link{GtkTreePath}}] A newly created \code{\link{GtkTreePath}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
