\alias{gtkTreePathNewFirst}
\name{gtkTreePathNewFirst}
\title{gtkTreePathNewFirst}
\description{Creates a new \code{\link{GtkTreePath}}.  The string representation of this path is "0"}
\usage{gtkTreePathNewFirst()}
\value{[\code{\link{GtkTreePath}}] A new \code{\link{GtkTreePath}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
