\alias{gtkTreePathNewFromIndices}
\name{gtkTreePathNewFromIndices}
\title{gtkTreePathNewFromIndices}
\description{Creates a new path with \code{first.index} and \code{varargs} as indices.}
\usage{gtkTreePathNewFromIndices(...)}
\arguments{\item{\verb{...}}{A newly created \code{\link{GtkTreePath}}.}}
\details{Since 2.2}
\value{[\code{\link{GtkTreePath}}] A newly created \code{\link{GtkTreePath}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
