\alias{gtkTreePathNext}
\name{gtkTreePathNext}
\title{gtkTreePathNext}
\description{Moves the \code{path} to point to the next node at the current depth.}
\usage{gtkTreePathNext(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreePath}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
