\alias{gtkTreePathPrependIndex}
\name{gtkTreePathPrependIndex}
\title{gtkTreePathPrependIndex}
\description{Prepends a new index to a path.  As a result, the depth of the path is
increased.}
\usage{gtkTreePathPrependIndex(object, index)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreePath}}]  A \code{\link{GtkTreePath}}.}
\item{\code{index}}{[integer]  The index.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
