\alias{gtkTreePathPrependIndex}
\name{gtkTreePathPrependIndex}
\title{gtkTreePathPrependIndex}
\description{Prepends a new index to a path.  As a result, the depth of the path is
increased.}
\usage{gtkTreePathPrependIndex(object, index)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreePath}}.}
\item{\verb{index}}{The index.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
