\alias{gtkTreeRowReferenceCopy}
\name{gtkTreeRowReferenceCopy}
\title{gtkTreeRowReferenceCopy}
\description{Copies a \verb{\link{GtkTreeRowReference}}.}
\usage{gtkTreeRowReferenceCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeRowReference}}]  a \verb{\link{GtkTreeRowReference}}}}
\details{  Since  2.2}
\value{[\verb{\link{GtkTreeRowReference}}]  a copy of \code{reference}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
