\alias{gtkTreeRowReferenceCopy}
\name{gtkTreeRowReferenceCopy}
\title{gtkTreeRowReferenceCopy}
\description{Copies a \code{\link{GtkTreeRowReference}}.}
\usage{gtkTreeRowReferenceCopy(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTreeRowReference}}}}
\details{Since 2.2}
\value{[\code{\link{GtkTreeRowReference}}] a copy of \code{reference}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
