\alias{gtkTreeRowReferenceGetModel}
\name{gtkTreeRowReferenceGetModel}
\title{gtkTreeRowReferenceGetModel}
\description{Returns the model which \code{reference} is monitoring in order to appropriately
the path.}
\usage{gtkTreeRowReferenceGetModel(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeRowReference}}]  A \code{\link{GtkTreeRowReference}}}}
\details{  Since  2.8}
\value{[\code{\link{GtkTreeModel}}]  The model, or \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
