\alias{gtkTreeRowReferenceGetModel}
\name{gtkTreeRowReferenceGetModel}
\title{gtkTreeRowReferenceGetModel}
\description{Returns the model that the row reference is monitoring.}
\usage{gtkTreeRowReferenceGetModel(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeRowReference}}]  A \code{\link{GtkTreeRowReference}}}}
\details{  Since  2.8}
\value{[\code{\link{GtkTreeModel}}]  the model}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
