\alias{gtkTreeRowReferenceGetPath}
\name{gtkTreeRowReferenceGetPath}
\title{gtkTreeRowReferenceGetPath}
\description{Returns a path that the row reference currently points to, or \code{NULL} if the
path pointed to is no longer valid.}
\usage{gtkTreeRowReferenceGetPath(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeRowReference}}}}
\value{[\code{\link{GtkTreePath}}] A current path, or \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
