\alias{gtkTreeRowReferenceInserted}
\name{gtkTreeRowReferenceInserted}
\title{gtkTreeRowReferenceInserted}
\description{Lets a set of row reference created by \code{\link{gtkTreeRowReferenceNewProxy}}
know that the model emitted the "row_inserted" signal.}
\usage{gtkTreeRowReferenceInserted(proxy, path)}
\arguments{
\item{\verb{proxy}}{A \code{\link{GObject}}}
\item{\verb{path}}{The row position that was inserted}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
