\alias{gtkTreeRowReferenceValid}
\name{gtkTreeRowReferenceValid}
\title{gtkTreeRowReferenceValid}
\description{Returns \code{TRUE} if the \code{reference} is non-\code{NULL} and refers to a current valid
path.}
\usage{gtkTreeRowReferenceValid(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeRowReference}}]  A \code{\link{GtkTreeRowReference}}, or \code{NULL}}}
\value{[logical]  \code{TRUE} if \code{reference} points to a valid path.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
