\alias{gtkTreeRowReferenceValid}
\name{gtkTreeRowReferenceValid}
\title{gtkTreeRowReferenceValid}
\description{Returns \code{TRUE} if the \code{reference} is non-\code{NULL} and refers to a current valid
path.}
\usage{gtkTreeRowReferenceValid(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeRowReference}}, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}}
\value{[logical] \code{TRUE} if \code{reference} points to a valid path.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
