\alias{gtkTreeSelectionGetTreeView}
\name{gtkTreeSelectionGetTreeView}
\title{gtkTreeSelectionGetTreeView}
\description{Returns the tree view associated with \code{selection}.}
\usage{gtkTreeSelectionGetTreeView(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeSelection}}]  A \verb{\link{GtkTreeSelection}}}}
\value{[\verb{\link{GtkTreeView}}]  A \verb{\link{GtkTreeView}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
