\alias{gtkTreeSelectionGetUserData}
\name{gtkTreeSelectionGetUserData}
\title{gtkTreeSelectionGetUserData}
\description{Returns the user data for the selection function.}
\usage{gtkTreeSelectionGetUserData(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeSelection}}]  A \code{\link{GtkTreeSelection}}.}}
\value{[R object]  The user data.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
