\alias{gtkTreeSelectionIterIsSelected}
\name{gtkTreeSelectionIterIsSelected}
\title{gtkTreeSelectionIterIsSelected}
\description{Returns \code{TRUE} if the row at \code{iter} is currently selected.}
\usage{gtkTreeSelectionIterIsSelected(object, iter)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeSelection}}]  A \code{\link{GtkTreeSelection}}}
\item{\code{iter}}{[\code{\link{GtkTreeIter}}]  A valid \code{\link{GtkTreeIter}}}
}
\value{[logical] \code{TRUE}, if \code{iter} is selected}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
