\alias{gtkTreeSelectionSelectAll}
\name{gtkTreeSelectionSelectAll}
\title{gtkTreeSelectionSelectAll}
\description{Selects all the nodes. \code{selection} must be set to \verb{GTK_SELECTION_MULTIPLE}
mode.}
\usage{gtkTreeSelectionSelectAll(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeSelection}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
