\alias{gtkTreeSelectionSelectIter}
\name{gtkTreeSelectionSelectIter}
\title{gtkTreeSelectionSelectIter}
\description{Selects the specified iterator.}
\usage{gtkTreeSelectionSelectIter(object, iter)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeSelection}}]  A \verb{\link{GtkTreeSelection}}.}
\item{\verb{iter}}{[\verb{\link{GtkTreeIter}}]  The \verb{\link{GtkTreeIter}} to be selected.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
