\alias{gtkTreeSelectionSelectIter}
\name{gtkTreeSelectionSelectIter}
\title{gtkTreeSelectionSelectIter}
\description{Selects the specified iterator.}
\usage{gtkTreeSelectionSelectIter(object, iter)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeSelection}}.}
\item{\verb{iter}}{The \code{\link{GtkTreeIter}} to be selected.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
