\alias{gtkTreeSelectionSelectPath}
\name{gtkTreeSelectionSelectPath}
\title{gtkTreeSelectionSelectPath}
\description{Select the row at \code{path}.}
\usage{gtkTreeSelectionSelectPath(object, path)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeSelection}}]  A \verb{\link{GtkTreeSelection}}.}
\item{\verb{path}}{[\verb{\link{GtkTreePath}}]  The \verb{\link{GtkTreePath}} to be selected.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
