\alias{gtkTreeSelectionSelectPath}
\name{gtkTreeSelectionSelectPath}
\title{gtkTreeSelectionSelectPath}
\description{Select the row at \code{path}.}
\usage{gtkTreeSelectionSelectPath(object, path)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeSelection}}.}
\item{\verb{path}}{The \code{\link{GtkTreePath}} to be selected.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
