\alias{gtkTreeSelectionUnselectAll}
\name{gtkTreeSelectionUnselectAll}
\title{gtkTreeSelectionUnselectAll}
\description{Unselects all the nodes.}
\usage{gtkTreeSelectionUnselectAll(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeSelection}}]  A \verb{\link{GtkTreeSelection}}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
