\alias{gtkTreeSelectionUnselectIter}
\name{gtkTreeSelectionUnselectIter}
\title{gtkTreeSelectionUnselectIter}
\description{Unselects the specified iterator.}
\usage{gtkTreeSelectionUnselectIter(object, iter)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeSelection}}]  A \code{\link{GtkTreeSelection}}.}
\item{\code{iter}}{[\code{\link{GtkTreeIter}}]  The \code{\link{GtkTreeIter}} to be unselected.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
