\alias{gtkTreeSelectionUnselectIter}
\name{gtkTreeSelectionUnselectIter}
\title{gtkTreeSelectionUnselectIter}
\description{Unselects the specified iterator.}
\usage{gtkTreeSelectionUnselectIter(object, iter)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeSelection}}.}
\item{\verb{iter}}{The \code{\link{GtkTreeIter}} to be unselected.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
