\alias{gtkTreeSelectionUnselectPath}
\name{gtkTreeSelectionUnselectPath}
\title{gtkTreeSelectionUnselectPath}
\description{Unselects the row at \code{path}.}
\usage{gtkTreeSelectionUnselectPath(object, path)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeSelection}}]  A \verb{\link{GtkTreeSelection}}.}
\item{\verb{path}}{[\verb{\link{GtkTreePath}}]  The \verb{\link{GtkTreePath}} to be unselected.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
