\alias{gtkTreeSelectionUnselectPath}
\name{gtkTreeSelectionUnselectPath}
\title{gtkTreeSelectionUnselectPath}
\description{Unselects the row at \code{path}.}
\usage{gtkTreeSelectionUnselectPath(object, path)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeSelection}}.}
\item{\verb{path}}{The \code{\link{GtkTreePath}} to be unselected.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
