\alias{gtkTreeSelectionUnselectRange}
\name{gtkTreeSelectionUnselectRange}
\title{gtkTreeSelectionUnselectRange}
\description{Unselects a range of nodes, determined by \code{start.path} and \code{end.path}
inclusive.}
\usage{gtkTreeSelectionUnselectRange(object, start.path, end.path)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeSelection}}.}
\item{\verb{start.path}}{The initial node of the range.}
\item{\verb{end.path}}{The initial node of the range.}
}
\details{Since 2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
