\alias{gtkTreeSortableSortColumnChanged}
\name{gtkTreeSortableSortColumnChanged}
\title{gtkTreeSortableSortColumnChanged}
\description{Emits a \code{\link{gtkTreeSortableSortColumnChanged}} signal on \code{sortable}.}
\usage{gtkTreeSortableSortColumnChanged(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeSortable}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
