\alias{gtkTreeStoreNewv}
\name{gtkTreeStoreNewv}
\title{gtkTreeStoreNewv}
\description{Non vararg creation function.  Used primarily by language bindings.}
\usage{gtkTreeStoreNewv(types)}
\arguments{\item{\verb{types}}{[\verb{\link{GType}}]  a list of \verb{\link{GType}} types for the columns, from first to last}}
\value{[\verb{\link{GtkTreeStore}}]  a new \verb{\link{GtkTreeStore}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
