\alias{gtkTreeStoreNewv}
\name{gtkTreeStoreNewv}
\title{gtkTreeStoreNewv}
\description{Non vararg creation function.  Used primarily by language bindings.}
\usage{gtkTreeStoreNewv(types)}
\arguments{\item{\verb{types}}{a list of \code{\link{GType}} types for the columns, from first to last. \emph{[  \acronym{array}   length=n_columns]}}}
\value{[\code{\link{GtkTreeStore}}] a new \code{\link{GtkTreeStore}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
