\alias{gtkTreeStoreSwap}
\name{gtkTreeStoreSwap}
\title{gtkTreeStoreSwap}
\description{Swaps \code{a} and \code{b} in the same level of \code{tree.store}. Note that this function
only works with unsorted stores.}
\usage{gtkTreeStoreSwap(object, a, b)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeStore}}]  A \code{\link{GtkTreeStore}}.}
\item{\code{a}}{[\code{\link{GtkTreeIter}}]  A \code{\link{GtkTreeIter}}.}
\item{\code{b}}{[\code{\link{GtkTreeIter}}]  Another \code{\link{GtkTreeIter}}.}
}
\details{  Since  2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
