\alias{gtkTreeStoreSwap}
\name{gtkTreeStoreSwap}
\title{gtkTreeStoreSwap}
\description{Swaps \code{a} and \code{b} in the same level of \code{tree.store}. Note that this function
only works with unsorted stores.}
\usage{gtkTreeStoreSwap(object, a, b)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeStore}}.}
\item{\verb{a}}{A \code{\link{GtkTreeIter}}.}
\item{\verb{b}}{Another \code{\link{GtkTreeIter}}.}
}
\details{Since 2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
