\alias{gtkTreeViewCollapseRow}
\name{gtkTreeViewCollapseRow}
\title{gtkTreeViewCollapseRow}
\description{Collapses a row (hides its child rows, if they exist).}
\usage{gtkTreeViewCollapseRow(object, path)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTreeView}}}
\item{\verb{path}}{path to a row in the \code{tree.view}}
}
\value{[logical] \code{TRUE} if the row was collapsed.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
