\alias{gtkTreeViewColumnClear}
\name{gtkTreeViewColumnClear}
\title{gtkTreeViewColumnClear}
\description{Unsets all the mappings on all renderers on the \code{tree.column}.}
\usage{gtkTreeViewColumnClear(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeViewColumn}}]  A \code{\link{GtkTreeViewColumn}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
