\alias{gtkTreeViewColumnClearAttributes}
\name{gtkTreeViewColumnClearAttributes}
\title{gtkTreeViewColumnClearAttributes}
\description{Clears all existing attributes previously set with
\code{\link{gtkTreeViewColumnSetAttributes}}.}
\usage{gtkTreeViewColumnClearAttributes(object, cell.renderer)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeViewColumn}}]  a \code{\link{GtkTreeViewColumn}}}
\item{\code{cell.renderer}}{[\code{\link{GtkCellRenderer}}]  a \code{\link{GtkCellRenderer}} to clear the attribute mapping on.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
