\alias{gtkTreeViewColumnClearAttributes}
\name{gtkTreeViewColumnClearAttributes}
\title{gtkTreeViewColumnClearAttributes}
\description{Clears all existing attributes previously set with
\code{\link{gtkTreeViewColumnSetAttributes}}.}
\usage{gtkTreeViewColumnClearAttributes(object, cell.renderer)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTreeViewColumn}}}
\item{\verb{cell.renderer}}{a \code{\link{GtkCellRenderer}} to clear the attribute mapping on.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
