\alias{gtkTreeViewColumnFocusCell}
\name{gtkTreeViewColumnFocusCell}
\title{gtkTreeViewColumnFocusCell}
\description{Sets the current keyboard focus to be at \code{cell}, if the column contains
2 or more editable and activatable cells.}
\usage{gtkTreeViewColumnFocusCell(object, cell)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeViewColumn}}}
\item{\verb{cell}}{A \code{\link{GtkCellRenderer}}}
}
\details{Since 2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
