\alias{gtkTreeViewColumnGetCellRenderers}
\name{gtkTreeViewColumnGetCellRenderers}
\title{gtkTreeViewColumnGetCellRenderers}
\description{Returns a newly-allocated \code{list} of all the cell renderers in the column, 
in no particular order.}
\usage{gtkTreeViewColumnGetCellRenderers(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeViewColumn}}]  A \code{\link{GtkTreeViewColumn}}}}
\value{[list]  A list of \code{GtkCellRenderers}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
