\alias{gtkTreeViewColumnGetCellRenderers}
\name{gtkTreeViewColumnGetCellRenderers}
\title{gtkTreeViewColumnGetCellRenderers}
\description{Returns a newly-allocated \verb{list} of all the cell renderers in the column, 
in no particular order.}
\usage{gtkTreeViewColumnGetCellRenderers(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeViewColumn}}]  A \verb{\link{GtkTreeViewColumn}}}}
\value{[list]  A list of \verb{GtkCellRenderers}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
