\alias{gtkTreeViewColumnGetExpand}
\name{gtkTreeViewColumnGetExpand}
\title{gtkTreeViewColumnGetExpand}
\description{Return \code{TRUE} if the column expands to take any available space.}
\usage{gtkTreeViewColumnGetExpand(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTreeViewColumn}}}}
\details{Since 2.4}
\value{[logical] \code{TRUE}, if the column expands}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
