\alias{gtkTreeViewColumnGetMaxWidth}
\name{gtkTreeViewColumnGetMaxWidth}
\title{gtkTreeViewColumnGetMaxWidth}
\description{Returns the maximum width in pixels of the \code{tree.column}, or -1 if no maximum
width is set.}
\usage{gtkTreeViewColumnGetMaxWidth(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeViewColumn}}.}}
\value{[integer] The maximum width of the \code{tree.column}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
