\alias{gtkTreeViewColumnGetReorderable}
\name{gtkTreeViewColumnGetReorderable}
\title{gtkTreeViewColumnGetReorderable}
\description{Returns \code{TRUE} if the \code{tree.column} can be reordered by the user.}
\usage{gtkTreeViewColumnGetReorderable(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeViewColumn}}}}
\value{[logical] \code{TRUE} if the \code{tree.column} can be reordered by the user.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
