\alias{gtkTreeViewColumnGetSizing}
\name{gtkTreeViewColumnGetSizing}
\title{gtkTreeViewColumnGetSizing}
\description{Returns the current type of \code{tree.column}.}
\usage{gtkTreeViewColumnGetSizing(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeViewColumn}}.}}
\value{[\code{\link{GtkTreeViewColumnSizing}}] The type of \code{tree.column}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
