\alias{gtkTreeViewColumnGetSortIndicator}
\name{gtkTreeViewColumnGetSortIndicator}
\title{gtkTreeViewColumnGetSortIndicator}
\description{Gets the value set by \code{\link{gtkTreeViewColumnSetSortIndicator}}.}
\usage{gtkTreeViewColumnGetSortIndicator(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTreeViewColumn}}}}
\value{[logical] whether the sort indicator arrow is displayed}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
