\alias{gtkTreeViewColumnGetSpacing}
\name{gtkTreeViewColumnGetSpacing}
\title{gtkTreeViewColumnGetSpacing}
\description{Returns the spacing of \code{tree.column}.}
\usage{gtkTreeViewColumnGetSpacing(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeViewColumn}}.}}
\value{[integer] the spacing of \code{tree.column}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
