\alias{gtkTreeViewColumnGetTitle}
\name{gtkTreeViewColumnGetTitle}
\title{gtkTreeViewColumnGetTitle}
\description{Returns the title of the widget.}
\usage{gtkTreeViewColumnGetTitle(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeViewColumn}}]  A \verb{\link{GtkTreeViewColumn}}.}}
\value{[character]  the title of the column.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
