\alias{gtkTreeViewColumnGetTitle}
\name{gtkTreeViewColumnGetTitle}
\title{gtkTreeViewColumnGetTitle}
\description{Returns the title of the widget.}
\usage{gtkTreeViewColumnGetTitle(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeViewColumn}}.}}
\value{[character] the title of the column. This string should not be
modified or freed.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
