\alias{gtkTreeViewColumnGetWidget}
\name{gtkTreeViewColumnGetWidget}
\title{gtkTreeViewColumnGetWidget}
\description{Returns the \code{\link{GtkWidget}} in the button on the column header.  If a custom
widget has not been set then \code{NULL} is returned.}
\usage{gtkTreeViewColumnGetWidget(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeViewColumn}}.}}
\value{[\code{\link{GtkWidget}}] The \code{\link{GtkWidget}} in the column header, or \code{NULL}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
