\alias{gtkTreeViewColumnNew}
\name{gtkTreeViewColumnNew}
\title{gtkTreeViewColumnNew}
\description{Creates a new \verb{\link{GtkTreeViewColumn}}.}
\usage{gtkTreeViewColumnNew()}
\value{[\verb{\link{GtkTreeViewColumn}}]  A newly created \verb{\link{GtkTreeViewColumn}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
