\alias{gtkTreeViewColumnSetAttributes}
\name{gtkTreeViewColumnSetAttributes}
\title{gtkTreeViewColumnSetAttributes}
\description{Sets the attributes in the list as the attributes of \code{tree.column}.
The attributes should be in attribute/column order, as in
\code{\link{gtkTreeViewColumnAddAttribute}}. All existing attributes
are removed, and replaced with the new attributes.}
\usage{gtkTreeViewColumnSetAttributes(object, cell.renderer, ...)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeViewColumn}}.}
\item{\verb{cell.renderer}}{the \code{\link{GtkCellRenderer}} we're setting the attributes of}
\item{\verb{...}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
