\alias{gtkTreeViewColumnSetSizing}
\name{gtkTreeViewColumnSetSizing}
\title{gtkTreeViewColumnSetSizing}
\description{Sets the growth behavior of \code{tree.column} to \code{type}.}
\usage{gtkTreeViewColumnSetSizing(object, type)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeViewColumn}}.}
\item{\verb{type}}{The \code{\link{GtkTreeViewColumnSizing}}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
