\alias{gtkTreeViewColumnSetSpacing}
\name{gtkTreeViewColumnSetSpacing}
\title{gtkTreeViewColumnSetSpacing}
\description{Sets the spacing field of \code{tree.column}, which is the number of pixels to
place between cell renderers packed into it.}
\usage{gtkTreeViewColumnSetSpacing(object, spacing)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeViewColumn}}]  A \verb{\link{GtkTreeViewColumn}}.}
\item{\verb{spacing}}{[integer]  distance between cell renderers in pixels.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
