\alias{gtkTreeViewColumnSetTitle}
\name{gtkTreeViewColumnSetTitle}
\title{gtkTreeViewColumnSetTitle}
\description{Sets the title of the \code{tree.column}.  If a custom widget has been set, then
this value is ignored.}
\usage{gtkTreeViewColumnSetTitle(object, title)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeViewColumn}}.}
\item{\verb{title}}{The title of the \code{tree.column}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
