\alias{gtkTreeViewColumnSetWidget}
\name{gtkTreeViewColumnSetWidget}
\title{gtkTreeViewColumnSetWidget}
\description{Sets the widget in the header to be \code{widget}.  If widget is \code{NULL}, then the
header button is set with a \code{\link{GtkLabel}} set to the title of \code{tree.column}.}
\usage{gtkTreeViewColumnSetWidget(object, widget = NULL)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeViewColumn}}.}
\item{\verb{widget}}{A child \code{\link{GtkWidget}}, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
