\alias{gtkTreeViewEnableModelDragDest}
\name{gtkTreeViewEnableModelDragDest}
\title{gtkTreeViewEnableModelDragDest}
\description{Turns \code{tree.view} into a drop destination for automatic DND.}
\usage{gtkTreeViewEnableModelDragDest(object, targets, actions)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeView}}]  a \code{\link{GtkTreeView}}}
\item{\code{targets}}{[\code{\link{GtkTargetEntry}}]  the table of targets that the drag will support}
\item{\code{actions}}{[\code{\link{GdkDragAction}}]  the bitmask of possible actions for a drag from this
   widget}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
