\alias{gtkTreeViewExpandToPath}
\name{gtkTreeViewExpandToPath}
\title{gtkTreeViewExpandToPath}
\description{Expands the row at \code{path}. This will also expand all parent rows of
\code{path} as necessary.}
\usage{gtkTreeViewExpandToPath(object, path)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeView}}.}
\item{\verb{path}}{path to a row.}
}
\details{Since 2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
