\alias{gtkTreeViewGetColumns}
\name{gtkTreeViewGetColumns}
\title{gtkTreeViewGetColumns}
\description{Returns a \code{list} of all the \code{\link{GtkTreeViewColumn}} s currently in \code{tree.view}.}
\usage{gtkTreeViewGetColumns(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeView}}]  A \code{\link{GtkTreeView}}}}
\value{[list]  A list of \code{\link{GtkTreeViewColumn}} s}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
