\alias{gtkTreeViewGetColumns}
\name{gtkTreeViewGetColumns}
\title{gtkTreeViewGetColumns}
\description{Returns a \verb{list} of all the \code{\link{GtkTreeViewColumn}} s currently in \code{tree.view}.}
\usage{gtkTreeViewGetColumns(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeView}}}}
\value{[list] A list of \code{\link{GtkTreeViewColumn}} s. \emph{[  \acronym{element-type}   GtkTreeViewColumn][  \acronym{transfer container}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
