\alias{gtkTreeViewGetModel}
\name{gtkTreeViewGetModel}
\title{gtkTreeViewGetModel}
\description{Returns the model the \code{\link{GtkTreeView}} is based on.  Returns \code{NULL} if the
model is unset.}
\usage{gtkTreeViewGetModel(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTreeView}}}}
\value{[\code{\link{GtkTreeModel}}] A \code{\link{GtkTreeModel}}, or \code{NULL} if none is currently being used.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
