\alias{gtkTreeViewGetRowSeparatorFunc}
\name{gtkTreeViewGetRowSeparatorFunc}
\title{gtkTreeViewGetRowSeparatorFunc}
\description{Returns the current row separator function.}
\usage{gtkTreeViewGetRowSeparatorFunc(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeView}}]  a \verb{\link{GtkTreeView}}}}
\details{  Since  2.6}
\value{[\verb{\link{GtkTreeViewRowSeparatorFunc}}]  the current row separator function.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
