\alias{gtkTreeViewGetSearchColumn}
\name{gtkTreeViewGetSearchColumn}
\title{gtkTreeViewGetSearchColumn}
\description{Gets the column searched on by the interactive search code.}
\usage{gtkTreeViewGetSearchColumn(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeView}}}}
\value{[integer] the column the interactive search code searches in.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
