\alias{gtkTreeViewGetSearchEqualFunc}
\name{gtkTreeViewGetSearchEqualFunc}
\title{gtkTreeViewGetSearchEqualFunc}
\description{Returns the compare function currently in use.}
\usage{gtkTreeViewGetSearchEqualFunc(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkTreeView}}]  A \verb{\link{GtkTreeView}}}}
\value{[\verb{\link{GtkTreeViewSearchEqualFunc}}]  the currently used compare function for the search code.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
