\alias{gtkTreeViewGetSearchEqualFunc}
\name{gtkTreeViewGetSearchEqualFunc}
\title{gtkTreeViewGetSearchEqualFunc}
\description{Returns the compare function currently in use.}
\usage{gtkTreeViewGetSearchEqualFunc(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeView}}}}
\value{[\code{\link{GtkTreeViewSearchEqualFunc}}] the currently used compare function for the search code.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
