\alias{gtkTreeViewGetSearchPositionFunc}
\name{gtkTreeViewGetSearchPositionFunc}
\title{gtkTreeViewGetSearchPositionFunc}
\description{Returns the positioning function currently in use.}
\usage{gtkTreeViewGetSearchPositionFunc(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeView}}}}
\details{Since 2.10}
\value{[\code{\link{GtkTreeViewSearchPositionFunc}}] the currently used function for positioning the search dialog.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
