\alias{gtkTreeViewGetSelection}
\name{gtkTreeViewGetSelection}
\title{gtkTreeViewGetSelection}
\description{Gets the \code{\link{GtkTreeSelection}} associated with \code{tree.view}.}
\usage{gtkTreeViewGetSelection(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeView}}.}}
\value{[\code{\link{GtkTreeSelection}}] A \code{\link{GtkTreeSelection}} object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
