\alias{gtkTreeViewGetShowExpanders}
\name{gtkTreeViewGetShowExpanders}
\title{gtkTreeViewGetShowExpanders}
\description{Returns whether or not expanders are drawn in \code{tree.view}.}
\usage{gtkTreeViewGetShowExpanders(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTreeView}}.}}
\details{Since 2.12}
\value{[logical] \code{TRUE} if expanders are drawn in \code{tree.view}, \code{FALSE}
otherwise.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
