\alias{gtkTreeViewGetVadjustment}
\name{gtkTreeViewGetVadjustment}
\title{gtkTreeViewGetVadjustment}
\description{Gets the \code{\link{GtkAdjustment}} currently being used for the vertical aspect.}
\usage{gtkTreeViewGetVadjustment(object)}
\arguments{\item{\verb{object}}{A \code{\link{GtkTreeView}}}}
\value{[\code{\link{GtkAdjustment}}] A \code{\link{GtkAdjustment}} object, or \code{NULL} if none is currently being
used.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
