\alias{gtkTreeViewMapExpandedRows}
\name{gtkTreeViewMapExpandedRows}
\title{gtkTreeViewMapExpandedRows}
\description{Calls \code{func} on all expanded rows.}
\usage{gtkTreeViewMapExpandedRows(object, func, data = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeView}}]  A \verb{\link{GtkTreeView}}}
\item{\verb{func}}{[\verb{\link{GtkTreeViewMappingFunc}}]  A function to be called}
\item{\verb{data}}{[R object]  User data to be passed to the function.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
