\alias{gtkTreeViewNew}
\name{gtkTreeViewNew}
\title{gtkTreeViewNew}
\description{Creates a new \verb{\link{GtkTreeView}} widget.}
\usage{gtkTreeViewNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}]  A newly created \verb{\link{GtkTreeView}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
