\alias{gtkTreeViewNew}
\name{gtkTreeViewNew}
\title{gtkTreeViewNew}
\description{Creates a new \code{\link{GtkTreeView}} widget.}
\usage{gtkTreeViewNew(show = TRUE)}
\value{[\code{\link{GtkWidget}}] A newly created \code{\link{GtkTreeView}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
