\alias{gtkTreeViewNewWithModel}
\name{gtkTreeViewNewWithModel}
\title{gtkTreeViewNewWithModel}
\description{Creates a new \verb{\link{GtkTreeView}} widget with the model initialized to \code{model}.}
\usage{gtkTreeViewNewWithModel(model = NULL, show = TRUE)}
\arguments{\item{\verb{model}}{[\verb{\link{GtkTreeModel}}]  the model.}}
\value{[\verb{\link{GtkWidget}}]  A newly created \verb{\link{GtkTreeView}} widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
