\alias{gtkTreeViewRemoveColumn}
\name{gtkTreeViewRemoveColumn}
\title{gtkTreeViewRemoveColumn}
\description{Removes \code{column} from \code{tree.view}.}
\usage{gtkTreeViewRemoveColumn(object, column)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeView}}]  A \code{\link{GtkTreeView}}.}
\item{\code{column}}{[\code{\link{GtkTreeViewColumn}}]  The \code{\link{GtkTreeViewColumn}} to remove.}
}
\value{[integer]  The number of columns in \code{tree.view} after removing.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
