\alias{gtkTreeViewRemoveColumn}
\name{gtkTreeViewRemoveColumn}
\title{gtkTreeViewRemoveColumn}
\description{Removes \code{column} from \code{tree.view}.}
\usage{gtkTreeViewRemoveColumn(object, column)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeView}}.}
\item{\verb{column}}{The \code{\link{GtkTreeViewColumn}} to remove.}
}
\value{[integer] The number of columns in \code{tree.view} after removing.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
