\alias{gtkTreeViewRowActivated}
\name{gtkTreeViewRowActivated}
\title{gtkTreeViewRowActivated}
\description{Activates the cell determined by \code{path} and \code{column}.}
\usage{gtkTreeViewRowActivated(object, path, column)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeView}}]  A \verb{\link{GtkTreeView}}}
\item{\verb{path}}{[\verb{\link{GtkTreePath}}]  The \verb{\link{GtkTreePath}} to be activated.}
\item{\verb{column}}{[\verb{\link{GtkTreeViewColumn}}]  The \verb{\link{GtkTreeViewColumn}} to be activated.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
