\alias{gtkTreeViewRowActivated}
\name{gtkTreeViewRowActivated}
\title{gtkTreeViewRowActivated}
\description{Activates the cell determined by \code{path} and \code{column}.}
\usage{gtkTreeViewRowActivated(object, path, column)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeView}}}
\item{\verb{path}}{The \code{\link{GtkTreePath}} to be activated.}
\item{\verb{column}}{The \code{\link{GtkTreeViewColumn}} to be activated.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
