\alias{gtkTreeViewRowExpanded}
\name{gtkTreeViewRowExpanded}
\title{gtkTreeViewRowExpanded}
\description{Returns \code{TRUE} if the node pointed to by \code{path} is expanded in \code{tree.view}.}
\usage{gtkTreeViewRowExpanded(object, path)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeView}}.}
\item{\verb{path}}{A \code{\link{GtkTreePath}} to test expansion state.}
}
\value{[logical] \code{TRUE} if \verb{path} is expanded.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
