\alias{gtkTreeViewSetDragDestRow}
\name{gtkTreeViewSetDragDestRow}
\title{gtkTreeViewSetDragDestRow}
\description{Sets the row that is highlighted for feedback.}
\usage{gtkTreeViewSetDragDestRow(object, path, pos)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeView}}]  a \verb{\link{GtkTreeView}}}
\item{\verb{path}}{[\verb{\link{GtkTreePath}}]  The path of the row to highlight, or \code{NULL}.}
\item{\verb{pos}}{[\code{\link{GtkTreeViewDropPosition}}]  Specifies whether to drop before, after or into the row}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
