\alias{gtkTreeViewSetDragDestRow}
\name{gtkTreeViewSetDragDestRow}
\title{gtkTreeViewSetDragDestRow}
\description{Sets the row that is highlighted for feedback.}
\usage{gtkTreeViewSetDragDestRow(object, path, pos)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTreeView}}}
\item{\verb{path}}{The path of the row to highlight, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
\item{\verb{pos}}{Specifies whether to drop before, after or into the row}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
