\alias{gtkTreeViewSetGridLines}
\name{gtkTreeViewSetGridLines}
\title{gtkTreeViewSetGridLines}
\description{Sets which grid lines to draw in \code{tree.view}.}
\usage{gtkTreeViewSetGridLines(object, grid.lines)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTreeView}}}
\item{\verb{grid.lines}}{a \code{\link{GtkTreeViewGridLines}} value indicating which grid lines to
enable.}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
