\alias{gtkTreeViewSetHadjustment}
\name{gtkTreeViewSetHadjustment}
\title{gtkTreeViewSetHadjustment}
\description{Sets the \code{\link{GtkAdjustment}} for the current horizontal aspect.}
\usage{gtkTreeViewSetHadjustment(object, adjustment)}
\arguments{
\item{\code{object}}{[\code{\link{GtkTreeView}}]  A \code{\link{GtkTreeView}}}
\item{\code{adjustment}}{[\code{\link{GtkAdjustment}}]  The \code{\link{GtkAdjustment}} to set, or \code{NULL}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
