\alias{gtkTreeViewSetHadjustment}
\name{gtkTreeViewSetHadjustment}
\title{gtkTreeViewSetHadjustment}
\description{Sets the \code{\link{GtkAdjustment}} for the current horizontal aspect.}
\usage{gtkTreeViewSetHadjustment(object, adjustment)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeView}}}
\item{\verb{adjustment}}{The \code{\link{GtkAdjustment}} to set, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
