\alias{gtkTreeViewSetHeadersClickable}
\name{gtkTreeViewSetHeadersClickable}
\title{gtkTreeViewSetHeadersClickable}
\description{Allow the column title buttons to be clicked.}
\usage{gtkTreeViewSetHeadersClickable(object, active)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeView}}.}
\item{\verb{active}}{\code{TRUE} if the columns are clickable.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
