\alias{gtkTreeViewSetHoverExpand}
\name{gtkTreeViewSetHoverExpand}
\title{gtkTreeViewSetHoverExpand}
\description{Enables of disables the hover expansion mode of \code{tree.view}.
Hover expansion makes rows expand or collaps if the pointer 
moves over them.}
\usage{gtkTreeViewSetHoverExpand(object, expand)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeView}}]  a \verb{\link{GtkTreeView}}}
\item{\verb{expand}}{[logical] \code{TRUE} to enable hover selection mode}
}
\details{  Since  2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
