\alias{gtkTreeViewSetHoverExpand}
\name{gtkTreeViewSetHoverExpand}
\title{gtkTreeViewSetHoverExpand}
\description{Enables of disables the hover expansion mode of \code{tree.view}.
Hover expansion makes rows expand or collapse if the pointer 
moves over them.}
\usage{gtkTreeViewSetHoverExpand(object, expand)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTreeView}}}
\item{\verb{expand}}{\code{TRUE} to enable hover selection mode}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
