\alias{gtkTreeViewSetHoverSelection}
\name{gtkTreeViewSetHoverSelection}
\title{gtkTreeViewSetHoverSelection}
\description{Enables of disables the hover selection mode of \code{tree.view}.
Hover selection makes the selected row follow the pointer.
Currently, this works only for the selection modes 
\code{GTK_SELECTION_SINGLE} and \code{GTK_SELECTION_BROWSE}.}
\usage{gtkTreeViewSetHoverSelection(object, hover)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTreeView}}}
\item{\verb{hover}}{\code{TRUE} to enable hover selection mode}
}
\details{Since 2.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
