\alias{gtkTreeViewSetRubberBanding}
\name{gtkTreeViewSetRubberBanding}
\title{gtkTreeViewSetRubberBanding}
\description{Enables or disables rubber banding in \code{tree.view}.  If the selection mode
is \verb{GTK_SELECTION_MULTIPLE}, rubber banding will allow the user to select
multiple rows by dragging the mouse.}
\usage{gtkTreeViewSetRubberBanding(object, enable)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTreeView}}}
\item{\verb{enable}}{\code{TRUE} to enable rubber banding}
}
\details{Since 2.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
