\alias{gtkTreeViewSetTooltipRow}
\name{gtkTreeViewSetTooltipRow}
\title{gtkTreeViewSetTooltipRow}
\description{Sets the tip area of \code{tooltip} to be the area covered by the row at \code{path}.
See also \code{\link{gtkTooltipSetTipArea}}.}
\usage{gtkTreeViewSetTooltipRow(object, tooltip, path)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkTreeView}}]  a \verb{\link{GtkTreeView}}}
\item{\verb{tooltip}}{[\verb{\link{GtkTooltip}}]  a \verb{\link{GtkTooltip}}}
\item{\verb{path}}{[\verb{\link{GtkTreePath}}]  a \verb{\link{GtkTreePath}}}
}
\details{  Since  2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
