\alias{gtkTreeViewSetTooltipRow}
\name{gtkTreeViewSetTooltipRow}
\title{gtkTreeViewSetTooltipRow}
\description{Sets the tip area of \code{tooltip} to be the area covered by the row at \code{path}.
See also \code{\link{gtkTreeViewSetTooltipColumn}} for a simpler alternative.
See also \code{\link{gtkTooltipSetTipArea}}.}
\usage{gtkTreeViewSetTooltipRow(object, tooltip, path)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkTreeView}}}
\item{\verb{tooltip}}{a \code{\link{GtkTooltip}}}
\item{\verb{path}}{a \code{\link{GtkTreePath}}}
}
\details{Since 2.12}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
