\alias{gtkTreeViewSetVadjustment}
\name{gtkTreeViewSetVadjustment}
\title{gtkTreeViewSetVadjustment}
\description{Sets the \code{\link{GtkAdjustment}} for the current vertical aspect.}
\usage{gtkTreeViewSetVadjustment(object, adjustment)}
\arguments{
\item{\verb{object}}{A \code{\link{GtkTreeView}}}
\item{\verb{adjustment}}{The \code{\link{GtkAdjustment}} to set, or \code{NULL}. \emph{[  \acronym{allow-none}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
