\alias{gtkTreeViewUnsetRowsDragDest}
\name{gtkTreeViewUnsetRowsDragDest}
\title{gtkTreeViewUnsetRowsDragDest}
\description{Undoes the effect of \code{\link{gtkTreeViewEnableModelDragDest}}.}
\usage{gtkTreeViewUnsetRowsDragDest(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkTreeView}}]  a \code{\link{GtkTreeView}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
