\alias{gtkTreeViewUnsetRowsDragDest}
\name{gtkTreeViewUnsetRowsDragDest}
\title{gtkTreeViewUnsetRowsDragDest}
\description{Undoes the effect of
\code{\link{gtkTreeViewEnableModelDragDest}}. Calling this method sets
\verb{"reorderable"} to \code{FALSE}.}
\usage{gtkTreeViewUnsetRowsDragDest(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTreeView}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
