\alias{gtkTreeViewUnsetRowsDragSource}
\name{gtkTreeViewUnsetRowsDragSource}
\title{gtkTreeViewUnsetRowsDragSource}
\description{Undoes the effect of
\code{\link{gtkTreeViewEnableModelDragSource}}. Calling this method sets
\verb{"reorderable"} to \code{FALSE}.}
\usage{gtkTreeViewUnsetRowsDragSource(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkTreeView}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
