\alias{gtkTrue}
\name{gtkTrue}
\title{gtkTrue}
\description{All this function does it to return \code{TRUE}. This can be useful for example
if you want to inhibit the deletion of a window. Of course you should
not do this as the user expects a reaction from clicking the close
icon of the window...}
\usage{gtkTrue()}
\value{[logical] \code{TRUE}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
