\alias{gtkUIManagerGetAccelGroup}
\name{gtkUIManagerGetAccelGroup}
\title{gtkUIManagerGetAccelGroup}
\description{Returns the \code{\link{GtkAccelGroup}} associated with \code{self}.}
\usage{gtkUIManagerGetAccelGroup(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkUIManager}} object}}
\details{Since 2.4}
\value{[\code{\link{GtkAccelGroup}}] the \code{\link{GtkAccelGroup}}. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
