\alias{gtkUIManagerGetActionGroups}
\name{gtkUIManagerGetActionGroups}
\title{gtkUIManagerGetActionGroups}
\description{Returns the list of action groups associated with \code{self}.}
\usage{gtkUIManagerGetActionGroups(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkUIManager}}]  a \code{\link{GtkUIManager}} object}}
\details{  Since  2.4}
\value{[list]  a \code{list} of action groups.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
