\alias{gtkUIManagerGetActionGroups}
\name{gtkUIManagerGetActionGroups}
\title{gtkUIManagerGetActionGroups}
\description{Returns the list of action groups associated with \code{self}.}
\usage{gtkUIManagerGetActionGroups(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkUIManager}} object}}
\details{Since 2.4}
\value{[list] a \verb{list} of
action groups. \emph{[  \acronym{element-type}   GtkActionGroup][  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
