\alias{gtkUIManagerGetAddTearoffs}
\name{gtkUIManagerGetAddTearoffs}
\title{gtkUIManagerGetAddTearoffs}
\description{Returns whether menus generated by this \code{\link{GtkUIManager}}
will have tearoff menu items.}
\usage{gtkUIManagerGetAddTearoffs(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkUIManager}}}}
\details{Since 2.4}
\value{[logical] whether tearoff menu items are added}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
