\alias{gtkUIManagerNewMergeId}
\name{gtkUIManagerNewMergeId}
\title{gtkUIManagerNewMergeId}
\description{Returns an unused merge id, suitable for use with 
\code{\link{gtkUIManagerAddUi}}.}
\usage{gtkUIManagerNewMergeId(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkUIManager}}}}
\details{Since 2.4}
\value{[numeric] an unused merge id.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
