\alias{gtkUIManagerRemoveActionGroup}
\name{gtkUIManagerRemoveActionGroup}
\title{gtkUIManagerRemoveActionGroup}
\description{Removes an action group from the list of action groups associated 
with \code{self}.}
\usage{gtkUIManagerRemoveActionGroup(object, action.group)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkUIManager}} object}
\item{\verb{action.group}}{the action group to be removed}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
