\alias{gtkUIManagerRemoveUi}
\name{gtkUIManagerRemoveUi}
\title{gtkUIManagerRemoveUi}
\description{Unmerges the part of \code{self}s content identified by \code{merge.id}.}
\usage{gtkUIManagerRemoveUi(object, merge.id)}
\arguments{
\item{\code{object}}{[\code{\link{GtkUIManager}}]  a \code{\link{GtkUIManager}} object}
\item{\code{merge.id}}{[numeric]  a merge id as returned by \code{\link{gtkUIManagerAddUiFromString}}}
}
\details{  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
