\alias{gtkUIManagerRemoveUi}
\name{gtkUIManagerRemoveUi}
\title{gtkUIManagerRemoveUi}
\description{Unmerges the part of \code{self}s content identified by \code{merge.id}.}
\usage{gtkUIManagerRemoveUi(object, merge.id)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkUIManager}} object}
\item{\verb{merge.id}}{a merge id as returned by \code{\link{gtkUIManagerAddUiFromString}}}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
