\alias{gtkUIManagerSetAddTearoffs}
\name{gtkUIManagerSetAddTearoffs}
\title{gtkUIManagerSetAddTearoffs}
\description{Sets the "add\_tearoffs" property, which controls whether menus 
generated by this \code{\link{GtkUIManager}} will have tearoff menu items. }
\usage{gtkUIManagerSetAddTearoffs(object, add.tearoffs)}
\arguments{
\item{\code{object}}{[\code{\link{GtkUIManager}}]  a \code{\link{GtkUIManager}}}
\item{\code{add.tearoffs}}{[logical]  whether tearoff menu items are added}
}
\details{Note that this only affects regular menus. Generated popup 
menus never have tearoff menu items.  

  Since  2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
