\alias{gtkVBoxNew}
\name{gtkVBoxNew}
\title{gtkVBoxNew}
\description{Creates a new GtkVBox.}
\usage{gtkVBoxNew(homogeneous = NULL, spacing = NULL, show = TRUE)}
\arguments{
\item{\verb{homogeneous}}{[logical] \code{TRUE} if all children are to be given equal space allotments.}
\item{\verb{spacing}}{[integer] the number of pixels to place by default between children.}
}
\value{[\verb{\link{GtkWidget}}] a new GtkVBox.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
