\alias{gtkVBoxNew}
\name{gtkVBoxNew}
\title{gtkVBoxNew}
\description{Creates a new \code{\link{GtkVBox}}.}
\usage{gtkVBoxNew(homogeneous = NULL, spacing = NULL, show = TRUE)}
\arguments{
\item{\verb{homogeneous}}{\code{TRUE} if all children are to be given equal space allotments.}
\item{\verb{spacing}}{the number of pixels to place by default between children.}
}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkVBox}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
