\alias{gtkVButtonBoxNew}
\name{gtkVButtonBoxNew}
\title{gtkVButtonBoxNew}
\description{Creates a new vertical button box.}
\usage{gtkVButtonBoxNew(show = TRUE)}
\value{[\verb{\link{GtkWidget}}] a new button box \verb{\link{GtkWidget}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
