\alias{gtkViewportGetHadjustment}
\name{gtkViewportGetHadjustment}
\title{gtkViewportGetHadjustment}
\description{Returns the horizontal adjustment of the viewport.}
\usage{gtkViewportGetHadjustment(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkViewport}}]  a \verb{\link{GtkViewport}}.}}
\value{[\verb{\link{GtkAdjustment}}]  the horizontal adjustment of \code{viewport}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
