\alias{gtkViewportGetShadowType}
\name{gtkViewportGetShadowType}
\title{gtkViewportGetShadowType}
\description{Gets the shadow type of the \code{\link{GtkViewport}}. See
\code{\link{gtkViewportSetShadowType}}.}
\usage{gtkViewportGetShadowType(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkViewport}}}}
\value{[\code{\link{GtkShadowType}}] the shadow type}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
