\alias{gtkViewportGetVadjustment}
\name{gtkViewportGetVadjustment}
\title{gtkViewportGetVadjustment}
\description{Returns the vertical adjustment of the viewport.}
\usage{gtkViewportGetVadjustment(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkViewport}}.}}
\value{[\code{\link{GtkAdjustment}}] the vertical adjustment of \code{viewport}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
