\alias{gtkViewportNew}
\name{gtkViewportNew}
\title{gtkViewportNew}
\description{Creates a new \code{\link{GtkViewport}} with the given adjustments.}
\usage{gtkViewportNew(hadjustment = NULL, vadjustment = NULL, show = TRUE)}
\arguments{
\item{\code{hadjustment}}{[\code{\link{GtkAdjustment}}]  horizontal adjustment.}
\item{\code{vadjustment}}{[\code{\link{GtkAdjustment}}]  vertical adjustment.}
}
\value{[\code{\link{GtkWidget}}]  a new \code{\link{GtkViewport}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
