\alias{gtkViewportSetHadjustment}
\name{gtkViewportSetHadjustment}
\title{gtkViewportSetHadjustment}
\description{Sets the horizontal adjustment of the viewport.}
\usage{gtkViewportSetHadjustment(object, adjustment = NULL)}
\arguments{
\item{\code{object}}{[\code{\link{GtkViewport}}]  a \code{\link{GtkViewport}}.}
\item{\code{adjustment}}{[\code{\link{GtkAdjustment}}]  a \code{\link{GtkAdjustment}}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
