\alias{gtkViewportSetHadjustment}
\name{gtkViewportSetHadjustment}
\title{gtkViewportSetHadjustment}
\description{Sets the horizontal adjustment of the viewport.}
\usage{gtkViewportSetHadjustment(object, adjustment = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkViewport}}]  a \verb{\link{GtkViewport}}.}
\item{\verb{adjustment}}{[\verb{\link{GtkAdjustment}}]  a \verb{\link{GtkAdjustment}}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
