\alias{gtkViewportSetHadjustment}
\name{gtkViewportSetHadjustment}
\title{gtkViewportSetHadjustment}
\description{Sets the horizontal adjustment of the viewport.}
\usage{gtkViewportSetHadjustment(object, adjustment = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkViewport}}.}
\item{\verb{adjustment}}{a \code{\link{GtkAdjustment}}. \emph{[  \acronym{allow-none}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
