\alias{gtkViewportSetShadowType}
\name{gtkViewportSetShadowType}
\title{gtkViewportSetShadowType}
\description{Sets the shadow type of the viewport.}
\usage{gtkViewportSetShadowType(object, type)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkViewport}}]  a \verb{\link{GtkViewport}}.}
\item{\verb{type}}{[\code{\link{GtkShadowType}}]  the new shadow type.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
