\alias{gtkViewportSetShadowType}
\name{gtkViewportSetShadowType}
\title{gtkViewportSetShadowType}
\description{Sets the shadow type of the viewport.}
\usage{gtkViewportSetShadowType(object, type)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkViewport}}.}
\item{\verb{type}}{the new shadow type.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
