\alias{gtkViewportSetVadjustment}
\name{gtkViewportSetVadjustment}
\title{gtkViewportSetVadjustment}
\description{Sets the vertical adjustment of the viewport.}
\usage{gtkViewportSetVadjustment(object, adjustment = NULL)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkViewport}}]  a \verb{\link{GtkViewport}}.}
\item{\verb{adjustment}}{[\verb{\link{GtkAdjustment}}]  a \verb{\link{GtkAdjustment}}.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
