\alias{gtkViewportSetVadjustment}
\name{gtkViewportSetVadjustment}
\title{gtkViewportSetVadjustment}
\description{Sets the vertical adjustment of the viewport.}
\usage{gtkViewportSetVadjustment(object, adjustment = NULL)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkViewport}}.}
\item{\verb{adjustment}}{a \code{\link{GtkAdjustment}}. \emph{[  \acronym{allow-none}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
