\alias{gtkVolumeButtonNew}
\name{gtkVolumeButtonNew}
\title{gtkVolumeButtonNew}
\description{Creates a \code{\link{GtkVolumeButton}}, with a range between 0.0 and 1.0, with
a stepping of 0.02. Volume values can be obtained and modified using
the functions from \code{\link{GtkScaleButton}}.}
\usage{gtkVolumeButtonNew(show = TRUE)}
\details{Since 2.12}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkVolumeButton}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
