\alias{gtkVScaleNew}
\name{gtkVScaleNew}
\title{gtkVScaleNew}
\description{Creates a new \verb{\link{GtkVScale}}.}
\usage{gtkVScaleNew(adjustment = NULL, show = TRUE)}
\arguments{\item{\verb{adjustment}}{[\verb{\link{GtkAdjustment}}] the \verb{\link{GtkAdjustment}} which sets the range of the scale.}}
\value{[\verb{\link{GtkWidget}}] a new \verb{\link{GtkVScale}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
