\alias{gtkVScaleNew}
\name{gtkVScaleNew}
\title{gtkVScaleNew}
\description{Creates a new \code{\link{GtkVScale}}.}
\usage{gtkVScaleNew(adjustment = NULL, show = TRUE)}
\arguments{\item{\verb{adjustment}}{the \code{\link{GtkAdjustment}} which sets the range of the scale.}}
\value{[\code{\link{GtkWidget}}] a new \code{\link{GtkVScale}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
