\alias{gtkVScrollbarNew}
\name{gtkVScrollbarNew}
\title{gtkVScrollbarNew}
\description{\emph{undocumented
}}
\usage{gtkVScrollbarNew(adjustment = NULL, show = TRUE)}
\arguments{\item{\code{adjustment}}{[\code{\link{GtkAdjustment}}] }}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
