\alias{gtkVScrollbarNew}
\name{gtkVScrollbarNew}
\title{gtkVScrollbarNew}
\description{Creates a new vertical scrollbar.}
\usage{gtkVScrollbarNew(adjustment = NULL, show = TRUE)}
\arguments{\item{\verb{adjustment}}{the \code{\link{GtkAdjustment}} to use, or \code{NULL} to create a new adjustment. \emph{[  \acronym{allow-none}  ]}}}
\value{[\code{\link{GtkWidget}}] the new \code{\link{GtkVScrollbar}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
