\alias{gtkWidgetAddEvents}
\name{gtkWidgetAddEvents}
\title{gtkWidgetAddEvents}
\description{Adds the events in the bitfield \code{events} to the event mask for
\code{widget}. See \code{\link{gtkWidgetSetEvents}} for details.}
\usage{gtkWidgetAddEvents(object, events)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}}
\item{\verb{events}}{[integer]  an event mask, see \code{\link{GdkEventMask}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
