\alias{gtkWidgetClassInstallStyleProperty}
\name{gtkWidgetClassInstallStyleProperty}
\title{gtkWidgetClassInstallStyleProperty}
\description{Installs a style property on a widget class. The parser for the
style property is determined by the value type of \code{pspec}.}
\usage{gtkWidgetClassInstallStyleProperty(klass, pspec)}
\arguments{
\item{\verb{klass}}{a \code{\link{GtkWidgetClass}}}
\item{\verb{pspec}}{the \code{\link{GParamSpec}} for the property}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
