\alias{gtkWidgetClassInstallStylePropertyParser}
\name{gtkWidgetClassInstallStylePropertyParser}
\title{gtkWidgetClassInstallStylePropertyParser}
\description{Installs a style property on a widget class.}
\usage{gtkWidgetClassInstallStylePropertyParser(klass, pspec, parser)}
\arguments{
\item{\verb{klass}}{[\verb{\link{GtkWidgetClass}}]  a \verb{\link{GtkWidgetClass}}}
\item{\verb{pspec}}{[\verb{\link{GParamSpec}}]  the \verb{\link{GParamSpec}} for the style property}
\item{\verb{parser}}{[GtkRcPropertyParser]  the parser for the style property}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
