\alias{gtkWidgetClassInstallStylePropertyParser}
\name{gtkWidgetClassInstallStylePropertyParser}
\title{gtkWidgetClassInstallStylePropertyParser}
\description{Installs a style property on a widget class.}
\usage{gtkWidgetClassInstallStylePropertyParser(klass, pspec, parser)}
\arguments{
\item{\verb{klass}}{a \code{\link{GtkWidgetClass}}}
\item{\verb{pspec}}{the \code{\link{GParamSpec}} for the style property}
\item{\verb{parser}}{the parser for the style property}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
