\alias{gtkWidgetCreatePangoContext}
\name{gtkWidgetCreatePangoContext}
\title{gtkWidgetCreatePangoContext}
\description{Creates a new \code{\link{PangoContext}} with the appropriate font map,
font description, and base direction for drawing text for
this widget. See also \code{\link{gtkWidgetGetPangoContext}}.}
\usage{gtkWidgetCreatePangoContext(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\value{[\code{\link{PangoContext}}] the new \code{\link{PangoContext}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
