\alias{gtkWidgetGetColormap}
\name{gtkWidgetGetColormap}
\title{gtkWidgetGetColormap}
\description{Gets the colormap that will be used to render \code{widget}. No reference will
be added to the returned colormap; it should not be unreferenced.}
\usage{gtkWidgetGetColormap(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\value{[\code{\link{GdkColormap}}] the colormap used by \code{widget}. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
