\alias{gtkWidgetGetDefaultColormap}
\name{gtkWidgetGetDefaultColormap}
\title{gtkWidgetGetDefaultColormap}
\description{Obtains the default colormap used to create widgets.}
\usage{gtkWidgetGetDefaultColormap()}
\value{[\verb{\link{GdkColormap}}]  default widget colormap}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
