\alias{gtkWidgetGetDefaultColormap}
\name{gtkWidgetGetDefaultColormap}
\title{gtkWidgetGetDefaultColormap}
\description{Obtains the default colormap used to create widgets.}
\usage{gtkWidgetGetDefaultColormap()}
\value{[\code{\link{GdkColormap}}] default widget colormap. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
