\alias{gtkWidgetGetDefaultStyle}
\name{gtkWidgetGetDefaultStyle}
\title{gtkWidgetGetDefaultStyle}
\description{Returns the default style used by all widgets initially.}
\usage{gtkWidgetGetDefaultStyle()}
\value{[\verb{\link{GtkStyle}}]  the default style.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
