\alias{gtkWidgetGetDefaultStyle}
\name{gtkWidgetGetDefaultStyle}
\title{gtkWidgetGetDefaultStyle}
\description{Returns the default style used by all widgets initially.}
\usage{gtkWidgetGetDefaultStyle()}
\value{[\code{\link{GtkStyle}}] the default style. This \code{\link{GtkStyle}} object is owned
by GTK+ and should not be modified or freed. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
