\alias{gtkWidgetGetDefaultVisual}
\name{gtkWidgetGetDefaultVisual}
\title{gtkWidgetGetDefaultVisual}
\description{Obtains the visual of the default colormap. Not really useful;
used to be useful before \code{\link{gdkColormapGetVisual}} existed.}
\usage{gtkWidgetGetDefaultVisual()}
\value{[\code{\link{GdkVisual}}] visual of the default colormap. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
