\alias{gtkWidgetGetExtensionEvents}
\name{gtkWidgetGetExtensionEvents}
\title{gtkWidgetGetExtensionEvents}
\description{Retrieves the extension events the widget will receive; see
\code{\link{gdkInputSetExtensionEvents}}.}
\usage{gtkWidgetGetExtensionEvents(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}}}
\value{[\code{\link{GdkExtensionMode}}]  extension events for \code{widget}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
