\alias{gtkWidgetGetName}
\name{gtkWidgetGetName}
\title{gtkWidgetGetName}
\description{Retrieves the name of a widget. See \code{\link{gtkWidgetSetName}} for the
significance of widget names.}
\usage{gtkWidgetGetName(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\value{[character] name of the widget.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
