\alias{gtkWidgetGetParent}
\name{gtkWidgetGetParent}
\title{gtkWidgetGetParent}
\description{Returns the parent container of \code{widget}.}
\usage{gtkWidgetGetParent(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}}}
\value{[\verb{\link{GtkWidget}}]  the parent container of \code{widget}, or \code{NULL}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
