\alias{gtkWidgetGetParent}
\name{gtkWidgetGetParent}
\title{gtkWidgetGetParent}
\description{Returns the parent container of \code{widget}.}
\usage{gtkWidgetGetParent(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\value{[\code{\link{GtkWidget}}] the parent container of \code{widget}, or \code{NULL}. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
