\alias{gtkWidgetGetParentWindow}
\name{gtkWidgetGetParentWindow}
\title{gtkWidgetGetParentWindow}
\description{Gets \code{widget}'s parent window.}
\usage{gtkWidgetGetParentWindow(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}.}}
\value{[\code{\link{GdkWindow}}] the parent window of \code{widget}. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
