\alias{gtkWidgetGetStyle}
\name{gtkWidgetGetStyle}
\title{gtkWidgetGetStyle}
\description{Simply an accessor function that returns \code{widget->style}.}
\usage{gtkWidgetGetStyle(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}}}
\value{[\verb{\link{GtkStyle}}]  the widget's \verb{\link{GtkStyle}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
