\alias{gtkWidgetGetStyle}
\name{gtkWidgetGetStyle}
\title{gtkWidgetGetStyle}
\description{Simply an accessor function that returns \code{widget->style}.}
\usage{gtkWidgetGetStyle(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\value{[\code{\link{GtkStyle}}] the widget's \code{\link{GtkStyle}}. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
