\alias{gtkWidgetGetTooltipWindow}
\name{gtkWidgetGetTooltipWindow}
\title{gtkWidgetGetTooltipWindow}
\description{Returns the \code{\link{GtkWindow}} of the current tooltip. This can be the
GtkWindow created by default, or the custom tooltip window set
using \code{\link{gtkWidgetSetTooltipWindow}}.}
\usage{gtkWidgetGetTooltipWindow(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.12}
\value{[\code{\link{GtkWindow}}] The \code{\link{GtkWindow}} of the current tooltip. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
