\alias{gtkWidgetGetVisual}
\name{gtkWidgetGetVisual}
\title{gtkWidgetGetVisual}
\description{Gets the visual that will be used to render \code{widget}.}
\usage{gtkWidgetGetVisual(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\value{[\code{\link{GdkVisual}}] the visual for \code{widget}. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
