\alias{gtkWidgetHideAll}
\name{gtkWidgetHideAll}
\title{gtkWidgetHideAll}
\description{Recursively hides a widget and any child widgets.}
\usage{gtkWidgetHideAll(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
