\alias{gtkWidgetIsSensitive}
\name{gtkWidgetIsSensitive}
\title{gtkWidgetIsSensitive}
\description{Returns the widget's effective sensitivity, which means
it is sensitive itself and also its parent widget is sensntive}
\usage{gtkWidgetIsSensitive(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\details{Since 2.18}
\value{[logical] \code{TRUE} if the widget is effectively sensitive}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
