\alias{gtkWidgetMap}
\name{gtkWidgetMap}
\title{gtkWidgetMap}
\description{This function is only for use in widget implementations. Causes
a widget to be mapped if it isn't already.}
\usage{gtkWidgetMap(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
