\alias{gtkWidgetMnemonicActivate}
\name{gtkWidgetMnemonicActivate}
\title{gtkWidgetMnemonicActivate}
\usage{gtkWidgetMnemonicActivate(object, group.cycling)}
\arguments{
\item{\code{object}}{[\code{\link{GtkWidget}}] }
\item{\code{group.cycling}}{[logical] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
