\alias{gtkWidgetMnemonicActivate}
\name{gtkWidgetMnemonicActivate}
\title{gtkWidgetMnemonicActivate}
\description{\emph{undocumented
}}
\usage{gtkWidgetMnemonicActivate(object, group.cycling)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkWidget}}] }
\item{\verb{group.cycling}}{[logical] }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
