\alias{gtkWidgetQueueClear}
\name{gtkWidgetQueueClear}
\title{gtkWidgetQueueClear}
\description{
This function does the same as \code{\link{gtkWidgetQueueDraw}}.
\strong{WARNING: \code{gtk_widget_queue_clear} is deprecated and should not be used in newly-written code.  
This function does the same as \code{\link{gtkWidgetQueueDraw}}.
  
\code{Deprecated}: Use \code{\link{gtkWidgetQueueDraw}} instead.  }
}
\usage{gtkWidgetQueueClear(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkWidget}}]  a \code{\link{GtkWidget}}}}
\details{\code{Deprecated}: Use \code{\link{gtkWidgetQueueDraw}} instead.  }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
