\alias{gtkWidgetQueueDraw}
\name{gtkWidgetQueueDraw}
\title{gtkWidgetQueueDraw}
\description{Equivalent to calling \code{\link{gtkWidgetQueueDrawArea}} for the
entire area of a widget.}
\usage{gtkWidgetQueueDraw(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
