\alias{gtkWidgetResetShapes}
\name{gtkWidgetResetShapes}
\title{gtkWidgetResetShapes}
\description{Recursively resets the shape on this widget and its descendants.}
\usage{gtkWidgetResetShapes(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkWidget}}]  a \verb{\link{GtkWidget}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
