\alias{gtkWidgetSet}
\name{gtkWidgetSet}
\title{gtkWidgetSet}
\description{
Precursor of \code{\link{gObjectSet}}.
\strong{WARNING: \code{gtk_widget_set} is deprecated and should not be used in newly-written code. Use \code{\link{gObjectSet}} instead.}
}
\usage{gtkWidgetSet(obj, ...)}
\arguments{\item{\code{...}}{ value of first property, followed by more properties, 
          \code{NULL}-terminated}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
