\alias{gtkWidgetSet}
\name{gtkWidgetSet}
\title{gtkWidgetSet}
\description{
Precursor of \code{\link{gObjectSet}}.
\strong{WARNING: \code{gtk_widget_set} has been deprecated since version 2.0 and should not be used in newly-written code. Use \code{\link{gObjectSet}} instead.}
}
\usage{gtkWidgetSet(obj, ...)}
\arguments{\item{\verb{...}}{\emph{undocumented
}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
