\alias{gtkWidgetSetCompositeName}
\name{gtkWidgetSetCompositeName}
\title{gtkWidgetSetCompositeName}
\description{Sets a widgets composite name. The widget must be
a composite child of its parent; see \code{\link{gtkWidgetPushCompositeChild}}.}
\usage{gtkWidgetSetCompositeName(object, name)}
\arguments{
\item{\code{object}}{[\code{\link{GtkWidget}}]  a \code{\link{GtkWidget}}.}
\item{\code{name}}{[character]  the name to set.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
