\alias{gtkWidgetSetDefaultColormap}
\name{gtkWidgetSetDefaultColormap}
\title{gtkWidgetSetDefaultColormap}
\description{Sets the default colormap to use when creating widgets.
\code{\link{gtkWidgetPushColormap}} is a better function to use if
you only want to affect a few widgets, rather than all widgets.}
\usage{gtkWidgetSetDefaultColormap(colormap)}
\arguments{\item{\code{colormap}}{[\code{\link{GdkColormap}}]  a \code{\link{GdkColormap}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
