\alias{gtkWidgetSetDefaultDirection}
\name{gtkWidgetSetDefaultDirection}
\title{gtkWidgetSetDefaultDirection}
\description{Sets the default reading direction for widgets where the
direction has not been explicitly set by \code{\link{gtkWidgetSetDirection}}.}
\usage{gtkWidgetSetDefaultDirection(dir)}
\arguments{\item{\verb{dir}}{the new default direction. This cannot be
\code{GTK_TEXT_DIR_NONE}.}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
