\alias{gtkWidgetSetParentWindow}
\name{gtkWidgetSetParentWindow}
\title{gtkWidgetSetParentWindow}
\description{Sets a non default parent window for \code{widget}.}
\usage{gtkWidgetSetParentWindow(object, parent.window)}
\arguments{
\item{\code{object}}{[\code{\link{GtkWidget}}]  a \code{\link{GtkWidget}}.}
\item{\code{parent.window}}{[\code{\link{GdkWindow}}]  the new parent window.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
