\alias{gtkWidgetSetParentWindow}
\name{gtkWidgetSetParentWindow}
\title{gtkWidgetSetParentWindow}
\description{Sets a non default parent window for \code{widget}.}
\usage{gtkWidgetSetParentWindow(object, parent.window)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWidget}}.}
\item{\verb{parent.window}}{the new parent window.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
