\alias{gtkWidgetSetState}
\name{gtkWidgetSetState}
\title{gtkWidgetSetState}
\description{This function is for use in widget implementations. Sets the state
of a widget (insensitive, prelighted, etc.) Usually you should set
the state using wrapper functions such as \code{\link{gtkWidgetSetSensitive}}.}
\usage{gtkWidgetSetState(object, state)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWidget}}}
\item{\verb{state}}{new state for \code{widget}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
