\alias{gtkWidgetShowAll}
\name{gtkWidgetShowAll}
\title{gtkWidgetShowAll}
\description{Recursively shows a widget, and any child widgets (if the widget is
a container).}
\usage{gtkWidgetShowAll(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWidget}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
