\alias{gtkWidgetSizeAllocate}
\name{gtkWidgetSizeAllocate}
\title{gtkWidgetSizeAllocate}
\description{This function is only used by \code{\link{GtkContainer}} subclasses, to assign a size
and position to their child widgets.}
\usage{gtkWidgetSizeAllocate(object, allocation)}
\arguments{
\item{\code{object}}{[\code{\link{GtkWidget}}]  a \code{\link{GtkWidget}}}
\item{\code{allocation}}{[\code{\link{GtkAllocation}}]  position and size to be allocated to \code{widget}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
