\alias{gtkWidgetSizeAllocate}
\name{gtkWidgetSizeAllocate}
\title{gtkWidgetSizeAllocate}
\description{This function is only used by \code{\link{GtkContainer}} subclasses, to assign a size
and position to their child widgets.}
\usage{gtkWidgetSizeAllocate(object, allocation)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWidget}}}
\item{\verb{allocation}}{position and size to be allocated to \code{widget}. \emph{[  \acronym{inout}  ]}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
