\alias{gtkWidgetStyleGet}
\name{gtkWidgetStyleGet}
\title{gtkWidgetStyleGet}
\description{Gets the values of a multiple style properties of \code{widget}.}
\usage{gtkWidgetStyleGet(object, ...)}
\arguments{\item{\code{...}}{[character]  the name of the first property to get}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
