\alias{gtkWidgetStyleGet}
\name{gtkWidgetStyleGet}
\title{gtkWidgetStyleGet}
\description{Gets the values of a multiple style properties of \code{widget}.}
\usage{gtkWidgetStyleGet(object, ...)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWidget}}}
\item{\verb{...}}{\emph{undocumented
}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
