\alias{gtkWidgetStyleGetProperty}
\name{gtkWidgetStyleGetProperty}
\title{gtkWidgetStyleGetProperty}
\description{Gets the value of a style property of \code{widget}.}
\usage{gtkWidgetStyleGetProperty(object, property.name)}
\arguments{
\item{\code{object}}{[\code{\link{GtkWidget}}]  a \code{\link{GtkWidget}}}
\item{\code{property.name}}{[character]  the name of a style property}
}
\value{
A list containing the following elements:
\item{\code{value}}{[R object]  location to return the property value }
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
