\alias{gtkWidgetStyleGetProperty}
\name{gtkWidgetStyleGetProperty}
\title{gtkWidgetStyleGetProperty}
\description{Gets the value of a style property of \code{widget}.}
\usage{gtkWidgetStyleGetProperty(object, property.name)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWidget}}}
\item{\verb{property.name}}{the name of a style property}
}
\value{
A list containing the following elements:
\item{\verb{value}}{location to return the property value}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
