\alias{gtkWindowAddMnemonic}
\name{gtkWindowAddMnemonic}
\title{gtkWindowAddMnemonic}
\description{Adds a mnemonic to this window.}
\usage{gtkWindowAddMnemonic(object, keyval, target)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWindow}}}
\item{\verb{keyval}}{the mnemonic}
\item{\verb{target}}{the widget that gets activated by the mnemonic}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
