\alias{gtkWindowGetDefaultIconList}
\name{gtkWindowGetDefaultIconList}
\title{gtkWindowGetDefaultIconList}
\description{Gets the value set by \code{\link{gtkWindowSetDefaultIconList}}.
but the pixbufs in the list have not had their reference count
incremented.}
\usage{gtkWindowGetDefaultIconList()}
\value{[list] copy of default icon list. \emph{[  \acronym{element-type}   GdkPixbuf][  \acronym{transfer container}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
