\alias{gtkWindowGetGravity}
\name{gtkWindowGetGravity}
\title{gtkWindowGetGravity}
\description{Gets the value set by \code{\link{gtkWindowSetGravity}}.}
\usage{gtkWindowGetGravity(object)}
\arguments{\item{\verb{object}}{[\verb{\link{GtkWindow}}]  a \verb{\link{GtkWindow}}}}
\value{[\code{\link{GdkGravity}}]  window gravity}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
