\alias{gtkWindowGetGravity}
\name{gtkWindowGetGravity}
\title{gtkWindowGetGravity}
\description{Gets the value set by \code{\link{gtkWindowSetGravity}}.}
\usage{gtkWindowGetGravity(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\value{[\code{\link{GdkGravity}}] window gravity. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
