\alias{gtkWindowGetIcon}
\name{gtkWindowGetIcon}
\title{gtkWindowGetIcon}
\description{Gets the value set by \code{\link{gtkWindowSetIcon}} (or if you've
called \code{\link{gtkWindowSetIconList}}, gets the first icon in
the icon list).}
\usage{gtkWindowGetIcon(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\value{[\code{\link{GdkPixbuf}}] icon for window. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
