\alias{gtkWindowGetModal}
\name{gtkWindowGetModal}
\title{gtkWindowGetModal}
\description{Returns whether the window is modal. See \code{\link{gtkWindowSetModal}}.}
\usage{gtkWindowGetModal(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\value{[logical] \code{TRUE} if the window is set to be modal and
establishes a grab when shown}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
