\alias{gtkWindowGetResizable}
\name{gtkWindowGetResizable}
\title{gtkWindowGetResizable}
\description{Gets the value set by \code{\link{gtkWindowSetResizable}}.}
\usage{gtkWindowGetResizable(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\value{[logical] \code{TRUE} if the user can resize the window}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
