\alias{gtkWindowGetScreen}
\name{gtkWindowGetScreen}
\title{gtkWindowGetScreen}
\description{Returns the \code{\link{GdkScreen}} associated with \code{window}.}
\usage{gtkWindowGetScreen(object)}
\arguments{\item{\code{object}}{[\code{\link{GtkWindow}}]  a \code{\link{GtkWindow}}.}}
\details{  Since  2.2}
\value{[\code{\link{GdkScreen}}]  a \code{\link{GdkScreen}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
