\alias{gtkWindowGetSkipTaskbarHint}
\name{gtkWindowGetSkipTaskbarHint}
\title{gtkWindowGetSkipTaskbarHint}
\description{Gets the value set by \code{\link{gtkWindowSetSkipTaskbarHint}}}
\usage{gtkWindowGetSkipTaskbarHint(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\details{Since 2.2}
\value{[logical] \code{TRUE} if window shouldn't be in taskbar}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
