\alias{gtkWindowGetTransientFor}
\name{gtkWindowGetTransientFor}
\title{gtkWindowGetTransientFor}
\description{Fetches the transient parent for this window. See
\code{\link{gtkWindowSetTransientFor}}.}
\usage{gtkWindowGetTransientFor(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\value{[\code{\link{GtkWindow}}] the transient parent for this window, or \code{NULL}
if no transient parent has been set. \emph{[  \acronym{transfer none}  ]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
