\alias{gtkWindowGetTypeHint}
\name{gtkWindowGetTypeHint}
\title{gtkWindowGetTypeHint}
\description{Gets the type hint for this window. See \code{\link{gtkWindowSetTypeHint}}.}
\usage{gtkWindowGetTypeHint(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\value{[\code{\link{GdkWindowTypeHint}}] the type hint for \code{window}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
