\alias{gtkWindowGetUrgencyHint}
\name{gtkWindowGetUrgencyHint}
\title{gtkWindowGetUrgencyHint}
\description{Gets the value set by \code{\link{gtkWindowSetUrgencyHint}}}
\usage{gtkWindowGetUrgencyHint(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\details{Since 2.8}
\value{[logical] \code{TRUE} if window is urgent}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
