\alias{GtkWindowGroup}
\alias{gtkWindowGroup}
\name{GtkWindowGroup}
\title{GtkWindowGroup}
\description{Limit the effect of grabs}
\section{Methods and Functions}{
\code{\link{gtkWindowGroupNew}()}\cr
\code{\link{gtkWindowGroupAddWindow}(object, window)}\cr
\code{\link{gtkWindowGroupRemoveWindow}(object, window)}\cr
\code{\link{gtkWindowGroupListWindows}(object)}\cr
\code{gtkWindowGroup()}
}
\section{Hierarchy}{\preformatted{GObject
   +----GtkWindowGroup}}
\section{Structures}{\describe{\item{\verb{GtkWindowGroup}}{
\emph{undocumented
}

}}}
\section{Convenient Construction}{\code{gtkWindowGroup} is the equivalent of \code{\link{gtkWindowGroupNew}}.}
\references{\url{https://developer-old.gnome.org/gtk2/stable/GtkWindowGroup.html}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
