\alias{gtkWindowGroupAddWindow}
\name{gtkWindowGroupAddWindow}
\title{gtkWindowGroupAddWindow}
\description{Adds a window to a \code{\link{GtkWindowGroup}}.}
\usage{gtkWindowGroupAddWindow(object, window)}
\arguments{
\item{\code{object}}{[\code{\link{GtkWindowGroup}}]  a \code{\link{GtkWindowGroup}}}
\item{\code{window}}{[\code{\link{GtkWindow}}]  the \code{\link{GtkWindow}} to add}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
