\alias{gtkWindowGroupAddWindow}
\name{gtkWindowGroupAddWindow}
\title{gtkWindowGroupAddWindow}
\description{Adds a window to a \verb{\link{GtkWindowGroup}}.}
\usage{gtkWindowGroupAddWindow(object, window)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkWindowGroup}}]  a \verb{\link{GtkWindowGroup}}}
\item{\verb{window}}{[\verb{\link{GtkWindow}}]  the \verb{\link{GtkWindow}} to add}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
