\alias{gtkWindowGroupAddWindow}
\name{gtkWindowGroupAddWindow}
\title{gtkWindowGroupAddWindow}
\description{Adds a window to a \code{\link{GtkWindowGroup}}.}
\usage{gtkWindowGroupAddWindow(object, window)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWindowGroup}}}
\item{\verb{window}}{the \code{\link{GtkWindow}} to add}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
