\alias{gtkWindowGroupNew}
\name{gtkWindowGroupNew}
\title{gtkWindowGroupNew}
\description{Creates a new \code{\link{GtkWindowGroup}} object. Grabs added with
\code{\link{gtkGrabAdd}} only affect windows within the same \code{\link{GtkWindowGroup}}.}
\usage{gtkWindowGroupNew()}
\value{[\code{\link{GtkWindowGroup}}] a new \code{\link{GtkWindowGroup}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
