\alias{gtkWindowGroupRemoveWindow}
\name{gtkWindowGroupRemoveWindow}
\title{gtkWindowGroupRemoveWindow}
\description{Removes a window from a \verb{\link{GtkWindowGroup}}.}
\usage{gtkWindowGroupRemoveWindow(object, window)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkWindowGroup}}]  a \verb{\link{GtkWindowGroup}}}
\item{\verb{window}}{[\verb{\link{GtkWindow}}]  the \verb{\link{GtkWindow}} to remove}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
