\alias{gtkWindowGroupRemoveWindow}
\name{gtkWindowGroupRemoveWindow}
\title{gtkWindowGroupRemoveWindow}
\description{Removes a window from a \code{\link{GtkWindowGroup}}.}
\usage{gtkWindowGroupRemoveWindow(object, window)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWindowGroup}}}
\item{\verb{window}}{the \code{\link{GtkWindow}} to remove}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
