\alias{gtkWindowMnemonicActivate}
\name{gtkWindowMnemonicActivate}
\title{gtkWindowMnemonicActivate}
\description{Activates the targets associated with the mnemonic.}
\usage{gtkWindowMnemonicActivate(object, keyval, modifier)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkWindow}}]  a \verb{\link{GtkWindow}}}
\item{\verb{keyval}}{[numeric]  the mnemonic}
\item{\verb{modifier}}{[\code{\link{GdkModifierType}}]  the modifiers }
}
\value{[logical] \code{TRUE} if the activation is done. }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
