\alias{gtkWindowMnemonicActivate}
\name{gtkWindowMnemonicActivate}
\title{gtkWindowMnemonicActivate}
\description{Activates the targets associated with the mnemonic.}
\usage{gtkWindowMnemonicActivate(object, keyval, modifier)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWindow}}}
\item{\verb{keyval}}{the mnemonic}
\item{\verb{modifier}}{the modifiers}
}
\value{[logical] \code{TRUE} if the activation is done.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
