\alias{gtkWindowRemoveAccelGroup}
\name{gtkWindowRemoveAccelGroup}
\title{gtkWindowRemoveAccelGroup}
\description{Reverses the effects of \code{\link{gtkWindowAddAccelGroup}}.}
\usage{gtkWindowRemoveAccelGroup(object, accel.group)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkWindow}}]  a \verb{\link{GtkWindow}}}
\item{\verb{accel.group}}{[\verb{\link{GtkAccelGroup}}]  a \verb{\link{GtkAccelGroup}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
