\alias{gtkWindowRemoveAccelGroup}
\name{gtkWindowRemoveAccelGroup}
\title{gtkWindowRemoveAccelGroup}
\description{Reverses the effects of \code{\link{gtkWindowAddAccelGroup}}.}
\usage{gtkWindowRemoveAccelGroup(object, accel.group)}
\arguments{
\item{\code{object}}{[\code{\link{GtkWindow}}]  a \code{\link{GtkWindow}}}
\item{\code{accel.group}}{[\code{\link{GtkAccelGroup}}]  a \code{\link{GtkAccelGroup}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
