\alias{gtkWindowRemoveAccelGroup}
\name{gtkWindowRemoveAccelGroup}
\title{gtkWindowRemoveAccelGroup}
\description{Reverses the effects of \code{\link{gtkWindowAddAccelGroup}}.}
\usage{gtkWindowRemoveAccelGroup(object, accel.group)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWindow}}}
\item{\verb{accel.group}}{a \code{\link{GtkAccelGroup}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
