\alias{gtkWindowRemoveMnemonic}
\name{gtkWindowRemoveMnemonic}
\title{gtkWindowRemoveMnemonic}
\description{Removes a mnemonic from this window.}
\usage{gtkWindowRemoveMnemonic(object, keyval, target)}
\arguments{
\item{\code{object}}{[\code{\link{GtkWindow}}]  a \code{\link{GtkWindow}}}
\item{\code{keyval}}{[numeric]  the mnemonic}
\item{\code{target}}{[\code{\link{GtkWidget}}]  the widget that gets activated by the mnemonic}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
