\alias{gtkWindowRemoveMnemonic}
\name{gtkWindowRemoveMnemonic}
\title{gtkWindowRemoveMnemonic}
\description{Removes a mnemonic from this window.}
\usage{gtkWindowRemoveMnemonic(object, keyval, target)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWindow}}}
\item{\verb{keyval}}{the mnemonic}
\item{\verb{target}}{the widget that gets activated by the mnemonic}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
