\alias{gtkWindowReshowWithInitialSize}
\name{gtkWindowReshowWithInitialSize}
\title{gtkWindowReshowWithInitialSize}
\description{Hides \code{window}, then reshows it, resetting the
default size and position of the window. Used
by GUI builders only.}
\usage{gtkWindowReshowWithInitialSize(object)}
\arguments{\item{\verb{object}}{a \code{\link{GtkWindow}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
