\alias{gtkWindowSetAcceptFocus}
\name{gtkWindowSetAcceptFocus}
\title{gtkWindowSetAcceptFocus}
\description{Windows may set a hint asking the desktop environment not to receive
the input focus. This function sets this hint.}
\usage{gtkWindowSetAcceptFocus(object, setting)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWindow}}}
\item{\verb{setting}}{\code{TRUE} to let this window receive input focus}
}
\details{Since 2.4}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
