\alias{gtkWindowSetMnemonicModifier}
\name{gtkWindowSetMnemonicModifier}
\title{gtkWindowSetMnemonicModifier}
\description{Sets the mnemonic modifier for this window.}
\usage{gtkWindowSetMnemonicModifier(object, modifier)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWindow}}}
\item{\verb{modifier}}{the modifier mask used to activate
mnemonics on this window.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
