\alias{gtkWindowSetResizable}
\name{gtkWindowSetResizable}
\title{gtkWindowSetResizable}
\description{Sets whether the user can resize a window. Windows are user resizable
by default.}
\usage{gtkWindowSetResizable(object, resizable)}
\arguments{
\item{\verb{object}}{[\verb{\link{GtkWindow}}]  a \verb{\link{GtkWindow}}}
\item{\verb{resizable}}{[logical] \code{TRUE} if the user can resize this window}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
