\alias{gtkWindowSetScreen}
\name{gtkWindowSetScreen}
\title{gtkWindowSetScreen}
\description{Sets the \code{\link{GdkScreen}} where the \code{window} is displayed; if
the window is already mapped, it will be unmapped, and
then remapped on the new screen.}
\usage{gtkWindowSetScreen(object, screen)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWindow}}.}
\item{\verb{screen}}{a \code{\link{GdkScreen}}.}
}
\details{Since 2.2}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
