\alias{gtkWindowSetUrgencyHint}
\name{gtkWindowSetUrgencyHint}
\title{gtkWindowSetUrgencyHint}
\description{Windows may set a hint asking the desktop environment to draw
the users attention to the window. This function sets this hint.}
\usage{gtkWindowSetUrgencyHint(object, setting)}
\arguments{
\item{\verb{object}}{a \code{\link{GtkWindow}}}
\item{\verb{setting}}{\code{TRUE} to mark this window as urgent}
}
\details{Since 2.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
