\alias{pangoAttrBackgroundNew}
\name{pangoAttrBackgroundNew}
\title{pangoAttrBackgroundNew}
\description{Create a new background color attribute.}
\usage{pangoAttrBackgroundNew(red, green, blue)}
\arguments{
\item{\verb{red}}{[integer]  the red value (ranging from 0 to 65535)}
\item{\verb{green}}{[integer]  the green value}
\item{\verb{blue}}{[integer]  the blue value}
}
\value{[\code{\link{PangoAttribute}}]  the newly allocated \code{\link{PangoAttribute}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
