\alias{pangoAttrFamilyNew}
\name{pangoAttrFamilyNew}
\title{pangoAttrFamilyNew}
\description{Create a new font family attribute.}
\usage{pangoAttrFamilyNew(family)}
\arguments{\item{\verb{family}}{[char]  the family or comma separated list of families}}
\value{[\verb{\link{PangoAttribute}}]  the newly allocated \verb{\link{PangoAttribute}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
