\alias{pangoAttrFamilyNew}
\name{pangoAttrFamilyNew}
\title{pangoAttrFamilyNew}
\description{Create a new font family attribute.}
\usage{pangoAttrFamilyNew(family)}
\arguments{\item{\code{family}}{[char]  the family or comma separated list of families}}
\value{[\code{\link{PangoAttribute}}]  the new \code{\link{PangoAttribute}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
