\alias{pangoAttrFontDescNew}
\name{pangoAttrFontDescNew}
\title{pangoAttrFontDescNew}
\description{Create a new font description attribute. (This attribute
allows setting family, style, weight, variant, stretch,
and size simultaneously.)}
\usage{pangoAttrFontDescNew(desc)}
\arguments{\item{\code{desc}}{[\code{\link{PangoFontDescription}}]  }}
\value{[\code{\link{PangoAttribute}}]  }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
