\alias{pangoAttrFontDescNew}
\name{pangoAttrFontDescNew}
\title{pangoAttrFontDescNew}
\description{Create a new font description attribute. This attribute
allows setting family, style, weight, variant, stretch,
and size simultaneously.}
\usage{pangoAttrFontDescNew(desc)}
\arguments{\item{\verb{desc}}{[\code{\link{PangoFontDescription}}]  the font description}}
\value{[\code{\link{PangoAttribute}}]   the newly allocated \code{\link{PangoAttribute}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
