\alias{pangoAttrForegroundNew}
\name{pangoAttrForegroundNew}
\title{pangoAttrForegroundNew}
\description{Create a new foreground color attribute.}
\usage{pangoAttrForegroundNew(red, green, blue)}
\arguments{
\item{\verb{red}}{the red value (ranging from 0 to 65535)}
\item{\verb{green}}{the green value}
\item{\verb{blue}}{the blue value}
}
\value{[\code{\link{PangoAttribute}}]  the newly allocated \code{\link{PangoAttribute}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
