\alias{pangoAttributeCopy}
\name{pangoAttributeCopy}
\title{pangoAttributeCopy}
\description{Make a copy of an attribute.}
\usage{pangoAttributeCopy(object)}
\arguments{\item{\verb{object}}{a \code{\link{PangoAttribute}}}}
\value{[\code{\link{PangoAttribute}}]  the newly allocated \code{\link{PangoAttribute}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
