\alias{pangoAttributeCopy}
\name{pangoAttributeCopy}
\title{pangoAttributeCopy}
\description{Make a copy of an attribute.}
\usage{pangoAttributeCopy(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoAttribute}}]  a \verb{\link{PangoAttribute}}}}
\value{[\verb{\link{PangoAttribute}}]  the newly allocated \verb{\link{PangoAttribute}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
