\alias{pangoAttributeInit}
\name{pangoAttributeInit}
\title{pangoAttributeInit}
\description{Initializes \code{attr}'s klass to \code{klass}, it's start\_index to zero,
and end\_index to \code{G_MAXUINT} such that the attribute applies
to the entire text by default.}
\usage{pangoAttributeInit(attr, klass)}
\arguments{
\item{\code{attr}}{[\code{\link{PangoAttribute}}]  a \code{\link{PangoAttribute}}}
\item{\code{klass}}{[\code{\link{PangoAttrClass}}]  a \code{PangoAttributeClass}}
}
\details{  Since  1.20}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
