\alias{pangoAttrIteratorCopy}
\name{pangoAttrIteratorCopy}
\title{pangoAttrIteratorCopy}
\description{Copy a \code{\link{PangoAttrIterator}}}
\usage{pangoAttrIteratorCopy(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoAttrIterator}}]  a \code{\link{PangoAttrIterator}}.}}
\value{[\code{\link{PangoAttrIterator}}]  the newly allocated \code{\link{PangoAttrIterator}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
