\alias{pangoAttrIteratorGetAttrs}
\name{pangoAttrIteratorGetAttrs}
\title{pangoAttrIteratorGetAttrs}
\description{Gets a list of all attributes at the current position of the
iterator.}
\usage{pangoAttrIteratorGetAttrs(object)}
\arguments{\item{\verb{object}}{a \code{\link{PangoAttrIterator}}}}
\details{Since 1.2}
\value{[list]  a list of
  all attributes for the current range.
  To free this value, \emph{[  \acronym{element-type}   Pango.Attribute): (transfer full]}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
