\alias{pangoAttrIteratorGetAttrs}
\name{pangoAttrIteratorGetAttrs}
\title{pangoAttrIteratorGetAttrs}
\description{Gets a list of all attributes at the current position of the
iterator.}
\usage{pangoAttrIteratorGetAttrs(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoAttrIterator}}]  a \code{\link{PangoAttrIterator}}}}
\details{  Since 1.2}
\value{[list] element-type Pango.Attribute): (transfer full. \acronym{element-type Pango.Attribute): (transfer} full. }
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
