\alias{pangoAttrLanguageNew}
\name{pangoAttrLanguageNew}
\title{pangoAttrLanguageNew}
\description{Create a new language tag attribute.}
\usage{pangoAttrLanguageNew(language)}
\arguments{\item{\verb{language}}{language tag}}
\value{[\code{\link{PangoAttribute}}] the newly allocated \code{\link{PangoAttribute}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
