\alias{pangoAttrLanguageNew}
\name{pangoAttrLanguageNew}
\title{pangoAttrLanguageNew}
\description{Create a new language tag attribute.}
\usage{pangoAttrLanguageNew(language)}
\arguments{\item{\verb{language}}{[\verb{\link{PangoLanguage}}]  language tag}}
\value{[\verb{\link{PangoAttribute}}]  the newly allocated \verb{\link{PangoAttribute}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
