\alias{pangoAttrListCopy}
\name{pangoAttrListCopy}
\title{pangoAttrListCopy}
\description{Copy \code{list} and return an identical new list.}
\usage{pangoAttrListCopy(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoAttrList}}]  a \code{\link{PangoAttrList}}}}
\value{[\code{\link{PangoAttrList}}]  the newly allocated \code{\link{PangoAttrList}}, with a
              reference count of one.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
