\alias{pangoAttrListGetIterator}
\name{pangoAttrListGetIterator}
\title{pangoAttrListGetIterator}
\description{Create a iterator initialized to the beginning of the list.}
\usage{pangoAttrListGetIterator(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoAttrList}}]  a \code{\link{PangoAttrList}}}}
\value{[\code{\link{PangoAttrIterator}}]  the newly allocated \code{\link{PangoAttrIterator}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
