\alias{pangoAttrListGetIterator}
\name{pangoAttrListGetIterator}
\title{pangoAttrListGetIterator}
\description{Create a iterator initialized to the beginning of the list.
\code{list} must not be modified until this iterator is freed.}
\usage{pangoAttrListGetIterator(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoAttrList}}]  a \code{\link{PangoAttrList}}}}
\value{[\code{\link{PangoAttrIterator}}]  the newly allocated \code{\link{PangoAttrIterator}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
