\alias{pangoAttrListNew}
\name{pangoAttrListNew}
\title{pangoAttrListNew}
\description{Create a new empty attribute list with a reference count of one.}
\usage{pangoAttrListNew()}
\value{[\verb{\link{PangoAttrList}}]  the newly allocated \verb{\link{PangoAttrList}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
