\alias{pangoAttrListNew}
\name{pangoAttrListNew}
\title{pangoAttrListNew}
\description{Create a new empty attribute list with a reference count of one.}
\usage{pangoAttrListNew()}
\value{[\code{\link{PangoAttrList}}]  the newly allocated \code{\link{PangoAttrList}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
