\alias{pangoAttrRiseNew}
\name{pangoAttrRiseNew}
\title{pangoAttrRiseNew}
\description{Create a new baseline displacement attribute.}
\usage{pangoAttrRiseNew(rise)}
\arguments{\item{\verb{rise}}{[integer]  the amount that the text should be displaced vertically,
       in Pango units. Positive values displace the text upwards.}}
\value{[\code{\link{PangoAttribute}}]  the newly allocated \code{\link{PangoAttribute}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
