\alias{pangoAttrSizeNew}
\name{pangoAttrSizeNew}
\title{pangoAttrSizeNew}
\description{Create a new font-size attribute in fractional points.}
\usage{pangoAttrSizeNew(size)}
\arguments{\item{\verb{size}}{[integer]  the font size, in \code{PANGO_SCALE}ths of a point.}}
\value{[\code{\link{PangoAttribute}}]  the newly allocated \code{\link{PangoAttribute}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
