\alias{pangoAttrSizeNewAbsolute}
\name{pangoAttrSizeNewAbsolute}
\title{pangoAttrSizeNewAbsolute}
\description{Create a new font-size attribute in device units.}
\usage{pangoAttrSizeNewAbsolute(size)}
\arguments{\item{\verb{size}}{[integer]  the font size, in \code{PANGO_SCALE}ths of a device unit.}}
\details{  Since 1.8}
\value{[\code{\link{PangoAttribute}}]  the newly allocated \code{\link{PangoAttribute}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
