\alias{pangoAttrStretchNew}
\name{pangoAttrStretchNew}
\title{pangoAttrStretchNew}
\description{Create a new font stretch attribute}
\usage{pangoAttrStretchNew(stretch)}
\arguments{\item{\verb{stretch}}{[\code{\link{PangoStretch}}]  the stretch}}
\value{[\code{\link{PangoAttribute}}]  the newly allocated \code{\link{PangoAttribute}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
