\alias{pangoAttrStrikethroughColorNew}
\name{pangoAttrStrikethroughColorNew}
\title{pangoAttrStrikethroughColorNew}
\description{Create a new strikethrough color attribute. This attribute
modifies the color of strikethrough lines. If not set, strikethrough
lines will use the foreground color.}
\usage{pangoAttrStrikethroughColorNew(red, green, blue)}
\arguments{
\item{\verb{red}}{[integer]  the red value (ranging from 0 to 65535)}
\item{\verb{green}}{[integer]  the green value}
\item{\verb{blue}}{[integer]  the blue value}
}
\details{  Since 1.8}
\value{[\code{\link{PangoAttribute}}]  the newly allocated \code{\link{PangoAttribute}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
