\alias{pangoAttrStrikethroughNew}
\name{pangoAttrStrikethroughNew}
\title{pangoAttrStrikethroughNew}
\description{Create a new font strike-through attribute.}
\usage{pangoAttrStrikethroughNew(strikethrough)}
\arguments{\item{\code{strikethrough}}{[logical]  \code{TRUE} if the text should be struck-through.}}
\value{[\code{\link{PangoAttribute}}]  the new \code{\link{PangoAttribute}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
