\alias{pangoAttrStrikethroughNew}
\name{pangoAttrStrikethroughNew}
\title{pangoAttrStrikethroughNew}
\description{Create a new strike-through attribute.}
\usage{pangoAttrStrikethroughNew(strikethrough)}
\arguments{\item{\verb{strikethrough}}{[logical] \code{TRUE} if the text should be struck-through.}}
\value{[\code{\link{PangoAttribute}}]  the newly allocated \code{\link{PangoAttribute}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
