\alias{pangoAttrStyleNew}
\name{pangoAttrStyleNew}
\title{pangoAttrStyleNew}
\description{Create a new font slant style attribute.}
\usage{pangoAttrStyleNew(style)}
\arguments{\item{\verb{style}}{[\code{\link{PangoStyle}}]  the slant style}}
\value{[\code{\link{PangoAttribute}}]  the newly allocated \code{\link{PangoAttribute}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
