\alias{pangoAttrTypeRegister}
\name{pangoAttrTypeRegister}
\title{pangoAttrTypeRegister}
\description{Allocate a new attribute type ID.}
\usage{pangoAttrTypeRegister(name)}
\arguments{\item{\verb{name}}{[character]  an identifier for the type (currently unused.)}}
\value{[\code{\link{PangoAttrType}}]  the new type ID.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
