\alias{pangoAttrTypeRegister}
\name{pangoAttrTypeRegister}
\title{pangoAttrTypeRegister}
\description{Allocate a new attribute type ID.  The attribute type name can be accessed
later by using \code{\link{pangoAttrTypeGetName}}.}
\usage{pangoAttrTypeRegister(name)}
\arguments{\item{\verb{name}}{[character]  an identifier for the type}}
\value{[\code{\link{PangoAttrType}}]  the new type ID.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
