\alias{pangoAttrUnderlineNew}
\name{pangoAttrUnderlineNew}
\title{pangoAttrUnderlineNew}
\description{Create a new underline-style attribute.}
\usage{pangoAttrUnderlineNew(underline)}
\arguments{\item{\verb{underline}}{[\code{\link{PangoUnderline}}]  the underline style.}}
\value{[\code{\link{PangoAttribute}}]  the newly allocated \code{\link{PangoAttribute}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
