\alias{pangoAttrWeightNew}
\name{pangoAttrWeightNew}
\title{pangoAttrWeightNew}
\description{Create a new font weight attribute.}
\usage{pangoAttrWeightNew(weight)}
\arguments{\item{\verb{weight}}{[\code{\link{PangoWeight}}]  the weight}}
\value{[\code{\link{PangoAttribute}}]  the newly allocated \code{\link{PangoAttribute}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
