\alias{pangoCairoContextGetFontOptions}
\name{pangoCairoContextGetFontOptions}
\title{pangoCairoContextGetFontOptions}
\description{\emph{undocumented
}}
\usage{pangoCairoContextGetFontOptions(context)}
\arguments{\item{\code{context}}{[\code{\link{PangoContext}}] }}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
