\alias{pangoCairoFontGetScaledFont}
\name{pangoCairoFontGetScaledFont}
\title{pangoCairoFontGetScaledFont}
\description{Gets the \code{\link{CairoScaledFont}} used by \code{font}.}
\usage{pangoCairoFontGetScaledFont(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoCairoFont}}]  a \code{\link{PangoFont}} from a \code{\link{PangoCairoFontMap}}}}
\details{  Since 1.18}
\value{[\code{\link{CairoScaledFont}}]  the \code{\link{CairoScaledFont}} used by \code{font},
              or \code{NULL} if \code{font} is \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
