\alias{pangoCairoFontMapCreateContext}
\name{pangoCairoFontMapCreateContext}
\title{pangoCairoFontMapCreateContext}
\description{Create a \code{\link{PangoContext}} for the given fontmap.}
\usage{pangoCairoFontMapCreateContext(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoCairoFontMap}}]  a \code{\link{PangoCairoFontMap}}}}
\details{  Since  1.10}
\value{[\code{\link{PangoContext}}] .}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
