\alias{pangoCairoFontMapCreateContext}
\name{pangoCairoFontMapCreateContext}
\title{pangoCairoFontMapCreateContext}
\description{
Create a \code{\link{PangoContext}} for the given fontmap.
\strong{WARNING: \code{pango_cairo_font_map_create_context} has been deprecated since version 1.22 and should not be used in newly-written code. Use \code{\link{pangoFontMapCreateContext}} instead.}
}
\usage{pangoCairoFontMapCreateContext(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoCairoFontMap}}]  a \code{\link{PangoCairoFontMap}}}}
\details{  Since 1.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
