\alias{pangoCairoFontMapGetDefault}
\name{pangoCairoFontMapGetDefault}
\title{pangoCairoFontMapGetDefault}
\description{Gets a default font map to use with Cairo.}
\usage{pangoCairoFontMapGetDefault()}
\details{  Since  1.10}
\value{[\code{\link{PangoFontMap}}]  the default Cairo fontmap for \code{\link{Pango}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
