\alias{pangoCairoFontMapGetFontType}
\name{pangoCairoFontMapGetFontType}
\title{pangoCairoFontMapGetFontType}
\description{Gets the type of Cairo font backend that \code{fontmap} uses.}
\usage{pangoCairoFontMapGetFontType(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoCairoFontMap}}]  a \code{\link{PangoCairoFontMap}}}}
\details{  Since 1.18}
\value{[\code{\link{CairoFontType}}]  the \code{\link{CairoFontType}} cairo font backend type}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
