\alias{pangoCairoFontMapGetResolution}
\name{pangoCairoFontMapGetResolution}
\title{pangoCairoFontMapGetResolution}
\description{Gets the resolution for the fontmap. See \code{\link{pangoCairoFontMapSetResolution}}}
\usage{pangoCairoFontMapGetResolution(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoCairoFontMap}}]  a \code{\link{PangoCairoFontMap}}}}
\details{  Since 1.10}
\value{[numeric]  the resolution in "dots per inch"}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
