\alias{pangoCairoShowLayout}
\name{pangoCairoShowLayout}
\title{pangoCairoShowLayout}
\description{Draws a \code{\link{PangoLayout}} in the specified cairo context.
The top-left corner of the \code{\link{PangoLayout}} will be drawn
at the current point of the cairo context.}
\usage{pangoCairoShowLayout(cr, layout)}
\arguments{
\item{\verb{cr}}{[\code{\link{Cairo}}]  a Cairo context}
\item{\verb{layout}}{[\code{\link{PangoLayout}}]  a Pango layout}
}
\details{  Since 1.10}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
