\alias{pangoColorFree}
\name{pangoColorFree}
\title{pangoColorFree}
\description{Frees a color allocated by \code{\link{pangoColorCopy}}.}
\usage{pangoColorFree(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoColor}}]  an allocated \verb{\link{PangoColor}}, may be \code{NULL}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
