\alias{pangoColorFree}
\name{pangoColorFree}
\title{pangoColorFree}
\description{Frees a color allocated by \code{\link{pangoColorCopy}}.}
\usage{pangoColorFree(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoColor}}]  an allocated \code{\link{PangoColor}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
