\alias{pangoContextGetBaseDir}
\name{pangoContextGetBaseDir}
\title{pangoContextGetBaseDir}
\description{Retrieves the base direction for the context. See
\code{\link{pangoContextSetBaseDir}}.}
\usage{pangoContextGetBaseDir(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoContext}}]  a \code{\link{PangoContext}}}}
\value{[\code{\link{PangoDirection}}]  the base direction for the context.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
