\alias{pangoContextGetFontDescription}
\name{pangoContextGetFontDescription}
\title{pangoContextGetFontDescription}
\description{Retrieve the default font description for the context.}
\usage{pangoContextGetFontDescription(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoContext}}]  a \code{\link{PangoContext}}}}
\value{[\code{\link{PangoFontDescription}}]  a pointer to the context's default font description.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
