\alias{pangoContextGetFontDescription}
\name{pangoContextGetFontDescription}
\title{pangoContextGetFontDescription}
\description{Retrieve the default font description for the context.}
\usage{pangoContextGetFontDescription(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoContext}}]  a \verb{\link{PangoContext}}}}
\value{[\verb{\link{PangoFontDescription}}]  a pointer to the context's default font description.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
