\alias{pangoContextGetFontMap}
\name{pangoContextGetFontMap}
\title{pangoContextGetFontMap}
\description{Gets the \verb{PangoFontmap} used to look up fonts for this context.}
\usage{pangoContextGetFontMap(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoContext}}]  a \code{\link{PangoContext}}}}
\details{  Since 1.6}
\value{[\code{\link{PangoFontMap}}]  the font map for the \code{\link{PangoContext}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
