\alias{pangoContextGetLanguage}
\name{pangoContextGetLanguage}
\title{pangoContextGetLanguage}
\description{Retrieves the global language tag for the context.}
\usage{pangoContextGetLanguage(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoContext}}]  a \code{\link{PangoContext}}}}
\value{[\code{\link{PangoLanguage}}]  the global language tag.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
