\alias{pangoContextSetFontDescription}
\name{pangoContextSetFontDescription}
\title{pangoContextSetFontDescription}
\description{Set the default font description for the context}
\usage{pangoContextSetFontDescription(object, desc)}
\arguments{
\item{\verb{object}}{[\verb{\link{PangoContext}}]  a \verb{\link{PangoContext}}}
\item{\verb{desc}}{[\verb{\link{PangoFontDescription}}]  the new pango font description}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
