\alias{pangoContextSetFontDescription}
\name{pangoContextSetFontDescription}
\title{pangoContextSetFontDescription}
\description{Set the default font description for the context}
\usage{pangoContextSetFontDescription(object, desc)}
\arguments{
\item{\code{object}}{[\code{\link{PangoContext}}]  a \code{\link{PangoContext}}}
\item{\code{desc}}{[\code{\link{PangoFontDescription}}]  the new pango font description}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
