\alias{pangoContextSetLanguage}
\name{pangoContextSetLanguage}
\title{pangoContextSetLanguage}
\description{Sets the global language tag for the context.}
\usage{pangoContextSetLanguage(object, language)}
\arguments{
\item{\code{object}}{[\code{\link{PangoContext}}]  a \code{\link{PangoContext}}}
\item{\code{language}}{[\code{\link{PangoLanguage}}]  the new language tag.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
