\alias{pangoContextSetLanguage}
\name{pangoContextSetLanguage}
\title{pangoContextSetLanguage}
\description{Sets the global language tag for the context.  The default language
for the locale of the running process can be found using
\code{pangoLanguageGetDefault()}.}
\usage{pangoContextSetLanguage(object, language)}
\arguments{
\item{\verb{object}}{[\verb{\link{PangoContext}}]  a \verb{\link{PangoContext}}}
\item{\verb{language}}{[\verb{\link{PangoLanguage}}]  the new language tag.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
