\alias{pangoCoverageGet}
\name{pangoCoverageGet}
\title{pangoCoverageGet}
\description{Determine whether a particular index is covered by \code{coverage}}
\usage{pangoCoverageGet(object, index)}
\arguments{
\item{\verb{object}}{[\code{\link{PangoCoverage}}]  a \code{\link{PangoCoverage}}}
\item{\verb{index}}{[integer]  the index to check}
}
\value{[\code{\link{PangoCoverageLevel}}]  the coverage level of \code{coverage} for character \code{index.}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
