\alias{pangoCoverageMax}
\name{pangoCoverageMax}
\title{pangoCoverageMax}
\description{Set the coverage for each index in \code{coverage} to be the max (better)
value of the current coverage for the index and the coverage for
the corresponding index in \code{other}.}
\usage{pangoCoverageMax(object, other)}
\arguments{
\item{\verb{object}}{[\code{\link{PangoCoverage}}]  a \code{\link{PangoCoverage}}}
\item{\verb{other}}{[\code{\link{PangoCoverage}}]  another \code{\link{PangoCoverage}}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
