\alias{pangoCoverageNew}
\name{pangoCoverageNew}
\title{pangoCoverageNew}
\description{Create a new \code{\link{PangoCoverage}}}
\usage{pangoCoverageNew()}
\value{[\code{\link{PangoCoverage}}]  a new PangoCoverage object, initialized to \code{PANGO_COVERAGE_NONE}
              with a reference count of 0.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
