\alias{pangoCoverageNew}
\name{pangoCoverageNew}
\title{pangoCoverageNew}
\description{Create a new \verb{\link{PangoCoverage}}}
\usage{pangoCoverageNew()}
\value{[\verb{\link{PangoCoverage}}]  the newly allocated \verb{\link{PangoCoverage}},
              initialized to \code{PANGO_COVERAGE_NONE}
              with a reference count of one.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
