\alias{pangoCoverageSet}
\name{pangoCoverageSet}
\title{pangoCoverageSet}
\description{Modify a particular index within \code{coverage}}
\usage{pangoCoverageSet(object, index, level)}
\arguments{
\item{\verb{object}}{a \code{\link{PangoCoverage}}}
\item{\verb{index}}{the index to modify}
\item{\verb{level}}{the new level for \code{index.}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
