\alias{pangoFontDescribe}
\name{pangoFontDescribe}
\title{pangoFontDescribe}
\description{Returns a description of the font, with font size set in points.
Use \code{\link{pangoFontDescribeWithAbsoluteSize}} if you want the font
size in device units.}
\usage{pangoFontDescribe(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoFont}}]  a \verb{\link{PangoFont}}}}
\value{[\verb{\link{PangoFontDescription}}]  a newly-allocated \verb{\link{PangoFontDescription}} object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
