\alias{pangoFontDescriptionCopy}
\name{pangoFontDescriptionCopy}
\title{pangoFontDescriptionCopy}
\description{Make a copy of a \code{\link{PangoFontDescription}}.}
\usage{pangoFontDescriptionCopy(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoFontDescription}}]  a \code{\link{PangoFontDescription}}}}
\value{[\code{\link{PangoFontDescription}}]  the newly allocated \code{\link{PangoFontDescription}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
