\alias{pangoFontDescriptionCopy}
\name{pangoFontDescriptionCopy}
\title{pangoFontDescriptionCopy}
\description{Make a copy of a \code{\link{PangoFontDescription}}.}
\usage{pangoFontDescriptionCopy(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoFontDescription}}]  a \code{\link{PangoFontDescription}}, may be \code{NULL}}}
\value{[\code{\link{PangoFontDescription}}]  the newly allocated \code{\link{PangoFontDescription}}, or \code{NULL}
              if \code{desc} was \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
