\alias{pangoFontDescriptionGetFamily}
\name{pangoFontDescriptionGetFamily}
\title{pangoFontDescriptionGetFamily}
\description{Gets the family name field of a font description. See
\code{\link{pangoFontDescriptionSetFamily}}.}
\usage{pangoFontDescriptionGetFamily(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoFontDescription}}]  a \code{\link{PangoFontDescription}}.}}
\value{[char]  The family name field. (Will be \code{NULL} if not previously set.)}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
