\alias{pangoFontDescriptionNew}
\name{pangoFontDescriptionNew}
\title{pangoFontDescriptionNew}
\description{Creates a new font description structure with all fields unset.}
\usage{pangoFontDescriptionNew()}
\value{[\code{\link{PangoFontDescription}}]  the newly-created \code{\link{PangoFontDescription}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
