\alias{pangoFontDescriptionUnsetFields}
\name{pangoFontDescriptionUnsetFields}
\title{pangoFontDescriptionUnsetFields}
\description{Unsets some of the fields in a \code{\link{PangoFontDescription}}.  The unset
fields will get back to their default values.}
\usage{pangoFontDescriptionUnsetFields(object, to.unset)}
\arguments{
\item{\verb{object}}{[\code{\link{PangoFontDescription}}]  a \code{\link{PangoFontDescription}}}
\item{\verb{to.unset}}{[\code{\link{PangoFontMask}}]  bitmask of fields in the \code{desc} to unset.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
