\alias{pangoFontFaceDescribe}
\name{pangoFontFaceDescribe}
\title{pangoFontFaceDescribe}
\description{Returns the family, style, variant, weight and stretch of
a \code{\link{PangoFontFace}}. The size field of the resulting font description
will be unset.}
\usage{pangoFontFaceDescribe(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoFontFace}}]  a \code{\link{PangoFontFace}}}}
\value{[\code{\link{PangoFontDescription}}]  a newly-created \code{\link{PangoFontDescription}} structure
 holding the description of the face.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
