\alias{pangoFontFaceIsSynthesized}
\name{pangoFontFaceIsSynthesized}
\title{pangoFontFaceIsSynthesized}
\description{Returns whether a \code{\link{PangoFontFace}} is synthesized by the underlying
font rendering engine from another face, perhaps by shearing, emboldening,
or lightening it.}
\usage{pangoFontFaceIsSynthesized(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoFontFace}}]  a \code{\link{PangoFontFace}}}}
\details{  Since 1.18}
\value{[logical]  whether \code{face} is synthesized.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
