\alias{pangoFontGetCoverage}
\name{pangoFontGetCoverage}
\title{pangoFontGetCoverage}
\description{Computes the coverage map for a given font and language tag.}
\usage{pangoFontGetCoverage(object, language)}
\arguments{
\item{\verb{object}}{a \code{\link{PangoFont}}}
\item{\verb{language}}{the language tag}
}
\value{[\code{\link{PangoCoverage}}] a newly-allocated \code{\link{PangoCoverage}} object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
