\alias{pangoFontGetFontMap}
\name{pangoFontGetFontMap}
\title{pangoFontGetFontMap}
\description{Gets the font map for which the font was created.}
\usage{pangoFontGetFontMap(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoFont}}]  a \code{\link{PangoFont}}}}
\details{  Since  1.10}
\value{[\code{\link{PangoFontMap}}]  the \code{\link{PangoFontMap}} for the font}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
