\alias{pangoFontsetGetMetrics}
\name{pangoFontsetGetMetrics}
\title{pangoFontsetGetMetrics}
\description{Get overall metric information for the fonts in the fontset.}
\usage{pangoFontsetGetMetrics(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoFontset}}]  a \verb{\link{PangoFontset}}}}
\value{[\verb{\link{PangoFontMetrics}}]  a \verb{\link{PangoFontMetrics}} object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
