\alias{pangoFontsetGetMetrics}
\name{pangoFontsetGetMetrics}
\title{pangoFontsetGetMetrics}
\description{Get overall metric information for the fonts in the fontset.}
\usage{pangoFontsetGetMetrics(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoFontset}}]  a \code{\link{PangoFontset}}}}
\value{[\code{\link{PangoFontMetrics}}]  a \code{\link{PangoFontMetrics}} object. The caller must call \code{pango\Font\Metrics\Unref()}
  when finished using the object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
