\alias{pangoFontsetGetMetrics}
\name{pangoFontsetGetMetrics}
\title{pangoFontsetGetMetrics}
\description{Get overall metric information for the fonts in the fontset.}
\usage{pangoFontsetGetMetrics(object)}
\arguments{\item{\verb{object}}{a \code{\link{PangoFontset}}}}
\value{[\code{\link{PangoFontMetrics}}] a \code{\link{PangoFontMetrics}} object.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
