\alias{pangoGlyphStringCopy}
\name{pangoGlyphStringCopy}
\title{pangoGlyphStringCopy}
\description{Copy a glyph string and associated storage.}
\usage{pangoGlyphStringCopy(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoGlyphString}}]  a PangoGlyphString.}}
\value{[\code{\link{PangoGlyphString}}] the copied PangoGlyphString}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
