\alias{pangoGlyphStringCopy}
\name{pangoGlyphStringCopy}
\title{pangoGlyphStringCopy}
\description{Copy a glyph string and associated storage.}
\usage{pangoGlyphStringCopy(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoGlyphString}}]  a \code{\link{PangoGlyphString}}, may be \code{NULL}}}
\value{[\code{\link{PangoGlyphString}}]  the newly allocated \code{\link{PangoGlyphString}},
              or \code{NULL} if \code{string} was \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
