\alias{pangoGlyphStringNew}
\name{pangoGlyphStringNew}
\title{pangoGlyphStringNew}
\description{Create a new \verb{\link{PangoGlyphString}}.}
\usage{pangoGlyphStringNew()}
\value{[\verb{\link{PangoGlyphString}}]  the newly allocated \verb{\link{PangoGlyphString}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
