\alias{pangoGlyphStringNew}
\name{pangoGlyphStringNew}
\title{pangoGlyphStringNew}
\description{Create a new \code{\link{PangoGlyphString}}.}
\usage{pangoGlyphStringNew()}
\value{[\code{\link{PangoGlyphString}}]  the newly allocated \code{\link{PangoGlyphString}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
