\alias{pangoGlyphStringSetSize}
\name{pangoGlyphStringSetSize}
\title{pangoGlyphStringSetSize}
\description{Resize a glyph string to the given length.}
\usage{pangoGlyphStringSetSize(object, new.len)}
\arguments{
\item{\verb{object}}{[\code{\link{PangoGlyphString}}]     a \code{\link{PangoGlyphString}}.}
\item{\verb{new.len}}{[integer]    the new length of the string.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
