\alias{pangoItemCopy}
\name{pangoItemCopy}
\title{pangoItemCopy}
\description{Copy an existing \code{\link{PangoItem}} structure.}
\usage{pangoItemCopy(item)}
\arguments{\item{\code{item}}{[\code{\link{PangoItem}}]  a \code{\link{PangoItem}}}}
\value{[\code{\link{PangoItem}}]  the new \code{\link{PangoItem}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
