\alias{pangoItemCopy}
\name{pangoItemCopy}
\title{pangoItemCopy}
\description{Copy an existing \code{\link{PangoItem}} structure.}
\usage{pangoItemCopy(item)}
\arguments{\item{\verb{item}}{[\code{\link{PangoItem}}]  a \code{\link{PangoItem}}, may be \code{NULL}}}
\value{[\code{\link{PangoItem}}]  the newly allocated \code{\link{PangoItem}}, or \code{NULL} if
              \code{item} was NULL.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
