\alias{pangoItemNew}
\name{pangoItemNew}
\title{pangoItemNew}
\description{Creates a new \verb{\link{PangoItem}} structure initialized to default values.}
\usage{pangoItemNew()}
\value{[\verb{\link{PangoItem}}]  the newly allocated \verb{\link{PangoItem}}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
