\alias{pangoLanguageToString}
\name{pangoLanguageToString}
\title{pangoLanguageToString}
\description{Gets the RFC-3066 format string representing the given language tag.}
\usage{pangoLanguageToString(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoLanguage}}]  a language tag.}}
\value{[char]  a string representing the language tag.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
