\alias{pangoLayoutCopy}
\name{pangoLayoutCopy}
\title{pangoLayoutCopy}
\description{Does a deep copy-by-value of the \code{src} layout. The attribute list,
tab list, and text from the original layout are all copied by
value.}
\usage{pangoLayoutCopy(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}}}
\value{[\code{\link{PangoLayout}}]  a new \code{\link{PangoLayout}} identical to \code{src}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
