\alias{pangoLayoutGetAlignment}
\name{pangoLayoutGetAlignment}
\title{pangoLayoutGetAlignment}
\description{Sets the alignment for the layout (how partial lines are
positioned within the horizontal space available.)}
\usage{pangoLayoutGetAlignment(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}}}
\value{[\code{\link{PangoAlignment}}]  the alignment value}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
