\alias{pangoLayoutGetIter}
\name{pangoLayoutGetIter}
\title{pangoLayoutGetIter}
\description{Returns an iterator to iterate over the visual extents of the layout.}
\usage{pangoLayoutGetIter(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoLayout}}]  a \verb{\link{PangoLayout}}}}
\value{[\verb{\link{PangoLayoutIter}}] .}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
