\alias{pangoLayoutGetIter}
\name{pangoLayoutGetIter}
\title{pangoLayoutGetIter}
\description{Returns an iterator to iterate over the visual extents of the layout.}
\usage{pangoLayoutGetIter(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}}}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
