\alias{pangoLayoutGetJustify}
\name{pangoLayoutGetJustify}
\title{pangoLayoutGetJustify}
\description{Gets whether or not each complete line should be stretched to
fill the entire width of the layout.}
\usage{pangoLayoutGetJustify(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}}}
\value{[logical]  the justify}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
