\alias{pangoLayoutGetLineCount}
\name{pangoLayoutGetLineCount}
\title{pangoLayoutGetLineCount}
\description{Retrieves the count of lines for the \code{layout}.}
\usage{pangoLayoutGetLineCount(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoLayout}}] \code{\link{PangoLayout}}}}
\value{[integer]  the line count.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
