\alias{pangoLayoutGetSpacing}
\name{pangoLayoutGetSpacing}
\title{pangoLayoutGetSpacing}
\description{Gets the amount of spacing between the lines of the layout.}
\usage{pangoLayoutGetSpacing(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}}}
\value{[integer]  the spacing (in \code{integer})}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
