\alias{pangoLayoutGetSpacing}
\name{pangoLayoutGetSpacing}
\title{pangoLayoutGetSpacing}
\description{Gets the amount of spacing between the lines of the layout.}
\usage{pangoLayoutGetSpacing(object)}
\arguments{\item{\verb{object}}{[\verb{\link{PangoLayout}}]  a \verb{\link{PangoLayout}}}}
\value{[integer]  the spacing in Pango units.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
