\alias{pangoLayoutGetWidth}
\name{pangoLayoutGetWidth}
\title{pangoLayoutGetWidth}
\description{Gets the width to which the lines of the \code{\link{PangoLayout}} should wrap.}
\usage{pangoLayoutGetWidth(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}}}
\value{[integer]  the width in Pango units, or -1 if no width set.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
