\alias{pangoLayoutGetWrap}
\name{pangoLayoutGetWrap}
\title{pangoLayoutGetWrap}
\description{Gets the wrap mode for the layout.}
\usage{pangoLayoutGetWrap(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}}}
\details{Use \code{\link{pangoLayoutIsWrapped}} to query whether any paragraphs
were actually wrapped.  }
\value{[\code{\link{PangoWrapMode}}]  active wrap mode.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
