\alias{pangoLayoutIterGetBaseline}
\name{pangoLayoutIterGetBaseline}
\title{pangoLayoutIterGetBaseline}
\description{Gets the Y position of the current line's baseline, in layout
coordinates (origin at top left of the entire layout).}
\usage{pangoLayoutIterGetBaseline(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoLayoutIter}}]  a \code{\link{PangoLayoutIter}}}}
\value{[integer]  baseline of current line.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
