\alias{pangoLayoutIterGetLayout}
\name{pangoLayoutIterGetLayout}
\title{pangoLayoutIterGetLayout}
\description{Gets the layout associated with a \code{\link{PangoLayoutIter}}.}
\usage{pangoLayoutIterGetLayout(iter)}
\arguments{\item{\verb{iter}}{[\code{\link{PangoLayoutIter}}]  a \code{\link{PangoLayoutIter}}}}
\details{  Since 1.20}
\value{[\code{\link{PangoLayout}}]  the layout associated with \code{iter}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
