\alias{pangoLayoutNew}
\name{pangoLayoutNew}
\title{pangoLayoutNew}
\description{Create a new \code{\link{PangoLayout}} object with attributes initialized to
default values for a particular \code{\link{PangoContext}}.}
\usage{pangoLayoutNew(context)}
\arguments{\item{\verb{context}}{[\code{\link{PangoContext}}]  a \code{\link{PangoContext}}}}
\value{[\code{\link{PangoLayout}}]  the newly allocated \code{\link{PangoLayout}}, with a reference
              count of one,}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
