\alias{pangoLayoutSetAttributes}
\name{pangoLayoutSetAttributes}
\title{pangoLayoutSetAttributes}
\description{Sets the text attributes for a layout object.
References \code{attrs}, so the caller can unref its reference.}
\usage{pangoLayoutSetAttributes(object, attrs)}
\arguments{
\item{\verb{object}}{[\verb{\link{PangoLayout}}]  a \verb{\link{PangoLayout}}}
\item{\verb{attrs}}{[\verb{\link{PangoAttrList}}]  a \verb{\link{PangoAttrList}}, can be \code{NULL}}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
