\alias{pangoLayoutSetWidth}
\name{pangoLayoutSetWidth}
\title{pangoLayoutSetWidth}
\description{Sets the width to which the lines of the \code{\link{PangoLayout}} should be wrapped.}
\usage{pangoLayoutSetWidth(object, width)}
\arguments{
\item{\code{object}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}.}
\item{\code{width}}{[integer]  the desired width, or -1 to indicate that no wrapping should be
        performed.}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
