\alias{pangoMatrixConcat}
\name{pangoMatrixConcat}
\title{pangoMatrixConcat}
\description{Changes the transformation represented by \code{matrix} to be the
transformation given by first applying transformation
given by \code{new.matrix} then applying the original transformation.}
\usage{pangoMatrixConcat(object, new.matrix)}
\arguments{
\item{\verb{object}}{[\code{\link{PangoMatrix}}]  a \code{\link{PangoMatrix}}}
\item{\verb{new.matrix}}{[\code{\link{PangoMatrix}}]  a \code{\link{PangoMatrix}}}
}
\details{  Since 1.6}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
