\alias{pangoMatrixCopy}
\name{pangoMatrixCopy}
\title{pangoMatrixCopy}
\description{Copies a \code{\link{PangoMatrix}}.}
\usage{pangoMatrixCopy(object)}
\arguments{\item{\code{object}}{[\code{\link{PangoMatrix}}]  a \code{\link{PangoMatrix}}}}
\details{  Since  1.6}
\value{[\code{\link{PangoMatrix}}]  a copy of \code{matrix}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
