\alias{pangoMatrixCopy}
\name{pangoMatrixCopy}
\title{pangoMatrixCopy}
\description{Copies a \code{\link{PangoMatrix}}.}
\usage{pangoMatrixCopy(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoMatrix}}]  a \code{\link{PangoMatrix}}, may be \code{NULL}}}
\details{  Since 1.6}
\value{[\code{\link{PangoMatrix}}]  the newly allocated \code{\link{PangoMatrix}}, or \code{NULL} if
              \code{matrix} was \code{NULL}.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
