\alias{pangoRendererDrawGlyph}
\name{pangoRendererDrawGlyph}
\title{pangoRendererDrawGlyph}
\description{Draws a single glyph with coordinates in device space.}
\usage{pangoRendererDrawGlyph(object, font, glyph, x, y)}
\arguments{
\item{\verb{object}}{a \code{\link{PangoRenderer}}}
\item{\verb{font}}{a \code{\link{PangoFont}}}
\item{\verb{glyph}}{the glyph index of a single glyph}
\item{\verb{x}}{X coordinate of left edge of baseline of glyph}
\item{\verb{y}}{Y coordinate of left edge of baseline of glyph}
}
\details{Since 1.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
