\alias{pangoRendererDrawGlyph}
\name{pangoRendererDrawGlyph}
\title{pangoRendererDrawGlyph}
\description{Draws a single glyph with coordinates in device space.}
\usage{pangoRendererDrawGlyph(object, font, glyph, x, y)}
\arguments{
\item{\code{object}}{[\code{\link{PangoRenderer}}]  a \code{\link{PangoRenderer}}}
\item{\code{font}}{[\code{\link{PangoFont}}]  a \code{\link{PangoFont}}}
\item{\code{glyph}}{[numeric]  the glyph index of a single glyph}
\item{\code{x}}{[numeric]  X coordinate of left edge of baseline of glyph}
\item{\code{y}}{[numeric]  Y coordinate of left edge of baseline of glyph}
}
\details{  Since  1.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
