\alias{pangoRendererDrawGlyphs}
\name{pangoRendererDrawGlyphs}
\title{pangoRendererDrawGlyphs}
\description{Draws the glyphs in \code{glyphs} with the specified \code{\link{PangoRenderer}}.}
\usage{pangoRendererDrawGlyphs(object, font, glyphs, x, y)}
\arguments{
\item{\verb{object}}{a \code{\link{PangoRenderer}}}
\item{\verb{font}}{a \code{\link{PangoFont}}}
\item{\verb{glyphs}}{a \code{\link{PangoGlyphString}}}
\item{\verb{x}}{X position of left edge of baseline, in user space coordinates
  in Pango units.}
\item{\verb{y}}{Y position of left edge of baseline, in user space coordinates
   in Pango units.}
}
\details{Since 1.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
