\alias{pangoRendererDrawGlyphs}
\name{pangoRendererDrawGlyphs}
\title{pangoRendererDrawGlyphs}
\description{Draws the glyphs in \code{glyphs} with the specified \code{\link{PangoRenderer}}.}
\usage{pangoRendererDrawGlyphs(object, font, glyphs, x, y)}
\arguments{
\item{\code{object}}{[\code{\link{PangoRenderer}}]  a \code{\link{PangoRenderer}}}
\item{\code{font}}{[\code{\link{PangoFont}}]  a \code{\link{PangoFont}}}
\item{\code{glyphs}}{[\code{\link{PangoGlyphString}}]  a \code{\link{PangoGlyphString}}}
\item{\code{x}}{[integer]  X position of left edge of baseline, in user space coordinates
  in Pango units.}
\item{\code{y}}{[integer]  Y position of left edge of baseline, in user space coordinates
   in Pango units.}
}
\details{  Since  1.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
