\alias{pangoRendererDrawLayout}
\name{pangoRendererDrawLayout}
\title{pangoRendererDrawLayout}
\description{Draws \code{layout} with the specified \code{\link{PangoRenderer}}.}
\usage{pangoRendererDrawLayout(object, layout, x, y)}
\arguments{
\item{\verb{object}}{a \code{\link{PangoRenderer}}}
\item{\verb{layout}}{a \code{\link{PangoLayout}}}
\item{\verb{x}}{X position of left edge of baseline, in user space coordinates
  in Pango units.}
\item{\verb{y}}{Y position of left edge of baseline, in user space coordinates
   in Pango units.}
}
\details{Since 1.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
