\alias{pangoRendererDrawLayout}
\name{pangoRendererDrawLayout}
\title{pangoRendererDrawLayout}
\description{Draws \code{layout} with the specified \code{\link{PangoRenderer}}.}
\usage{pangoRendererDrawLayout(object, layout, x, y)}
\arguments{
\item{\code{object}}{[\code{\link{PangoRenderer}}]  a \code{\link{PangoRenderer}}}
\item{\code{layout}}{[\code{\link{PangoLayout}}]  a \code{\link{PangoLayout}}}
\item{\code{x}}{[integer]  x position of left edge of baseline, in user space coordinates
  in Pango units.}
\item{\code{y}}{[integer]  x position of left edge of baseline, in user space coordinates
   in Pango units.}
}
\details{  Since  1.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
