\alias{pangoRendererDrawLayoutLine}
\name{pangoRendererDrawLayoutLine}
\title{pangoRendererDrawLayoutLine}
\description{Draws \code{line} with the specified \code{\link{PangoRenderer}}.}
\usage{pangoRendererDrawLayoutLine(object, line, x, y)}
\arguments{
\item{\verb{object}}{a \code{\link{PangoRenderer}}}
\item{\verb{line}}{a \code{\link{PangoLayoutLine}}}
\item{\verb{x}}{X position of left edge of baseline, in user space coordinates
  in Pango units.}
\item{\verb{y}}{Y position of left edge of baseline, in user space coordinates
   in Pango units.}
}
\details{Since 1.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
