\alias{pangoRendererDrawLayoutLine}
\name{pangoRendererDrawLayoutLine}
\title{pangoRendererDrawLayoutLine}
\description{Draws \code{line} with the specified \code{\link{PangoRenderer}}.}
\usage{pangoRendererDrawLayoutLine(object, line, x, y)}
\arguments{
\item{\code{object}}{[\code{\link{PangoRenderer}}]  a \code{\link{PangoRenderer}}}
\item{\code{line}}{[\code{\link{PangoLayoutLine}}]  a \code{\link{PangoLayoutLine}}}
\item{\code{x}}{[integer]  x position of left edge of baseline, in user space coordinates
  in Pango units.}
\item{\code{y}}{[integer]  x position of left edge of baseline, in user space coordinates
   in Pango units.}
}
\details{  Since  1.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
