\alias{pangoRendererSetColor}
\name{pangoRendererSetColor}
\title{pangoRendererSetColor}
\description{Sets the color for part of the rendering.}
\usage{pangoRendererSetColor(object, part, color)}
\arguments{
\item{\code{object}}{[\code{\link{PangoRenderer}}]  a \code{\link{PangoRenderer}}}
\item{\code{part}}{[\code{\link{PangoRenderPart}}]  the part to change the color of}
\item{\code{color}}{[\code{\link{PangoColor}}]  the new color or \code{NULL} to unset the current color}
}
\details{  Since  1.8}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
