\alias{pangoScriptIterNext}
\name{pangoScriptIterNext}
\title{pangoScriptIterNext}
\description{Advances a \code{\link{PangoScriptIter}} to the next range. If \code{iter}
is already at the end, it is left unchanged and \code{FALSE}
is returned.}
\usage{pangoScriptIterNext(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoScriptIter}}]  a \code{\link{PangoScriptIter}}}}
\details{  Since 1.4}
\value{[logical] \code{TRUE} if \code{iter} was successfully advanced.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
