\alias{pangoTabArrayCopy}
\name{pangoTabArrayCopy}
\title{pangoTabArrayCopy}
\description{Copies a \code{\link{PangoTabArray}}}
\usage{pangoTabArrayCopy(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoTabArray}}] \code{\link{PangoTabArray}} to copy}}
\value{[\code{\link{PangoTabArray}}]  the newly allocated \code{\link{PangoTabArray}},}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
