\alias{pangoTabArrayGetPositionsInPixels}
\name{pangoTabArrayGetPositionsInPixels}
\title{pangoTabArrayGetPositionsInPixels}
\description{Returns \code{TRUE} if the tab positions are in pixels, \code{FALSE} if they are
in Pango units.}
\usage{pangoTabArrayGetPositionsInPixels(object)}
\arguments{\item{\verb{object}}{[\code{\link{PangoTabArray}}]  a \code{\link{PangoTabArray}}}}
\value{[logical]  whether positions are in pixels.}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
