\alias{pangoTabArrayResize}
\name{pangoTabArrayResize}
\title{pangoTabArrayResize}
\description{Resizes a tab list. You must subsequently initialize any tabs that
were added as a result of growing the list.}
\usage{pangoTabArrayResize(object, new.size)}
\arguments{
\item{\verb{object}}{[\code{\link{PangoTabArray}}]  a \code{\link{PangoTabArray}}}
\item{\verb{new.size}}{[integer]  new size of the list}
}
\author{Derived by RGtkGen from GTK+ documentation}
\keyword{internal}
