\alias{atkActionGetType}
\alias{atkComponentGetType}
\alias{atkDocumentGetType}
\alias{atkEditableTextGetType}
\alias{atkGObjectAccessibleGetType}
\alias{atkHyperlinkGetType}
\alias{atkHypertextGetType}
\alias{atkImageGetType}
\alias{atkImplementorGetType}
\alias{atkKeyEventStructGetType}
\alias{atkNoOpObjectFactoryGetType}
\alias{atkNoOpObjectGetType}
\alias{atkObjectFactoryGetType}
\alias{atkObjectGetType}
\alias{atkRegistryGetType}
\alias{atkRelationGetType}
\alias{atkRelationSetGetType}
\alias{atkSelectionGetType}
\alias{atkStateSetGetType}
\alias{atkStreamableContentGetType}
\alias{atkTableGetType}
\alias{atkTextGetType}
\alias{atkUtilGetType}
\alias{atkValueGetType}
\alias{gdkColormapGetType}
\alias{gdkCursorGetType}
\alias{gdkDeviceGetType}
\alias{gdkDisplayGetType}
\alias{gdkDisplayManagerGetType}
\alias{gdkDragContextGetType}
\alias{gdkDrawableGetType}
\alias{gdkEventAnyGetType}
\alias{gdkEventGetType}
\alias{gdkFontGetType}
\alias{gdkGCGetType}
\alias{gdkImageGetType}
\alias{gdkKeymapGetType}
\alias{gdkPangoRendererGetType}
\alias{gdkPixbufAnimationGetType}
\alias{gdkPixbufAnimationIterGetType}
\alias{gdkPixbufLoaderGetType}
\alias{gdkPixmapGetType}
\alias{gdkScreenGetType}
\alias{gdkVisualGetType}
\alias{gdkWindowObjectGetType}
\alias{gladeXMLGetType}
\alias{gtkAboutDialogGetType}
\alias{gtkAccelGroupGetType}
\alias{gtkAccelLabelGetType}
\alias{gtkAccelMapGetType}
\alias{gtkAccessibleGetType}
\alias{gtkActionGetType}
\alias{gtkActionGroupGetType}
\alias{gtkAdjustmentGetType}
\alias{gtkAlignmentGetType}
\alias{gtkArrowGetType}
\alias{gtkAspectFrameGetType}
\alias{gtkAssistantGetType}
\alias{gtkAssistantPageTypeGetType}
\alias{gtkCellRendererAccelGetType}
\alias{gtkCellRendererAccelModeGetType}
\alias{gtkCellRendererSpinGetType} 
\alias{gtkLinkButtonGetType}
\alias{gtkBinGetType}
\alias{gtkBorderGetType}
\alias{gtkBoxGetType}
\alias{gtkButtonBoxGetType}
\alias{gtkButtonGetType}
\alias{gtkCListGetType}
\alias{gtkCTreeGetType}
\alias{gtkCTreeNodeGetType}
\alias{gtkCalendarGetType}
\alias{gtkCellEditableGetType}
\alias{gtkCellLayoutGetType}
\alias{gtkCellRendererComboGetType}
\alias{gtkCellRendererGetType}
\alias{gtkCellRendererPixbufGetType}
\alias{gtkCellRendererProgressGetType}
\alias{gtkCellRendererTextGetType}
\alias{gtkCellRendererToggleGetType}
\alias{gtkCellViewGetType}
\alias{gtkCheckButtonGetType}
\alias{gtkCheckMenuItemGetType}
\alias{gtkClipboardGetType}
\alias{gtkColorButtonGetType}
\alias{gtkColorSelectionDialogGetType}
\alias{gtkColorSelectionGetType}
\alias{gtkComboBoxEntryGetType}
\alias{gtkComboBoxGetType}
\alias{gtkComboGetType}
\alias{gtkContainerGetType}
\alias{gtkCurveGetType}
\alias{gtkDialogGetType}
\alias{gtkDrawingAreaGetType}
\alias{gtkEditableGetType}
\alias{gtkEntryCompletionGetType}
\alias{gtkEntryGetType}
\alias{gtkEventBoxGetType}
\alias{gtkExpanderGetType}
\alias{gtkFileChooserButtonGetType}
\alias{gtkFileChooserDialogGetType}
\alias{gtkFileChooserGetType}
\alias{gtkFileChooserWidgetGetType}
\alias{gtkFileFilterGetType}
\alias{gtkFileSelectionGetType}
\alias{gtkFixedGetType}
\alias{gtkFontButtonGetType}
\alias{gtkFontSelectionDialogGetType}
\alias{gtkFontSelectionGetType}
\alias{gtkFrameGetType}
\alias{gtkGammaCurveGetType}
\alias{gtkHBoxGetType}
\alias{gtkHButtonBoxGetType}
\alias{gtkHPanedGetType}
\alias{gtkHRulerGetType}
\alias{gtkHScaleGetType}
\alias{gtkHScrollbarGetType}
\alias{gtkHSeparatorGetType}
\alias{gtkHandleBoxGetType}
\alias{gtkIMContextGetType}
\alias{gtkIMContextSimpleGetType}
\alias{gtkIMMulticontextGetType}
\alias{gtkIconFactoryGetType}
\alias{gtkIconInfoGetType}
\alias{gtkIconSetGetType}
\alias{gtkIconSourceGetType}
\alias{gtkIconThemeGetType}
\alias{gtkIconViewGetType}
\alias{gtkImageGetType}
\alias{gtkImageMenuItemGetType}
\alias{gtkInputDialogGetType}
\alias{gtkInvisibleGetType}
\alias{gtkItemFactoryGetType}
\alias{gtkItemGetType}
\alias{gtkLabelGetType}
\alias{gtkLayoutGetType}
\alias{gtkListGetType}
\alias{gtkListItemGetType}
\alias{gtkListStoreGetType}
\alias{gtkMenuBarGetType}
\alias{gtkMenuGetType}
\alias{gtkMenuItemGetType}
\alias{gtkMenuShellGetType}
\alias{gtkMenuToolButtonGetType}
\alias{gtkMessageDialogGetType}
\alias{gtkMiscGetType}
\alias{gtkNotebookGetType}
\alias{gtkOldEditableGetType}
\alias{gtkOptionMenuGetType}
\alias{gtkPanedGetType}
\alias{gtkPixmapGetType}
\alias{gtkPlugGetType}
\alias{gtkPreviewGetType}
\alias{gtkPageOrientationGetType}
\alias{gtkPageSetGetType}
\alias{gtkPageSetupGetType}
\alias{gtkPaperSizeGetType}
\alias{gtkPrintContextGetType}
\alias{gtkPrintDuplexGetType} 
\alias{gtkPrintOperationActionGetType} 
\alias{gtkPrintOperationGetType}
\alias{gtkPrintOperationPreviewGetType}
\alias{gtkPrintPagesGetType}
\alias{gtkPrintQualityGetType}
\alias{gtkPrintSettingsGetType}
\alias{gtkRecentChooserDialogGetType}
\alias{gtkRecentChooserErrorGetType}
\alias{gtkRecentChooserGetType}
\alias{gtkRecentChooserMenuGetType}
\alias{gtkRecentChooserWidgetGetType}
\alias{gtkRecentFilterFlagsGetType}
\alias{gtkRecentFilterGetType}
\alias{gtkRecentInfoGetType} 
\alias{gtkRecentManagerErrorGetType}
\alias{gtkRecentManagerGetType}
\alias{gtkRecentSortTypeGetType} 
\alias{gtkSensitivityTypeGetType} 
\alias{gtkStatusIconGetType}
\alias{gtkTargetListGetType}
\alias{gtkTextBufferTargetInfoGetType} 
\alias{gtkProgressBarGetType}
\alias{gtkProgressGetType}
\alias{gtkRadioActionGetType}
\alias{gtkRadioButtonGetType}
\alias{gtkRadioMenuItemGetType}
\alias{gtkRadioToolButtonGetType}
\alias{gtkRangeGetType}
\alias{gtkRcStyleGetType}
\alias{gtkRequisitionGetType}
\alias{gtkRulerGetType}
\alias{gtkScaleGetType}
\alias{gtkScrollbarGetType}
\alias{gtkScrolledWindowGetType}
\alias{gtkSelectionDataGetType}
\alias{gtkSeparatorGetType}
\alias{gtkSeparatorMenuItemGetType}
\alias{gtkSeparatorToolItemGetType}
\alias{gtkSettingsGetType}
\alias{gtkSizeGroupGetType}
\alias{gtkSocketGetType}
\alias{gtkSpinButtonGetType}
\alias{gtkStatusbarGetType}
\alias{gtkStyleGetType}
\alias{gtkTableGetType}
\alias{gtkTearoffMenuItemGetType}
\alias{gtkTextAttributesGetType}
\alias{gtkTextBufferGetType}
\alias{gtkTextChildAnchorGetType}
\alias{gtkTextIterGetType}
\alias{gtkTextMarkGetType}
\alias{gtkTextTagGetType}
\alias{gtkTextTagTableGetType}
\alias{gtkTextViewGetType}
\alias{gtkTipsQueryGetType}
\alias{gtkToggleActionGetType}
\alias{gtkToggleButtonGetType}
\alias{gtkToggleToolButtonGetType}
\alias{gtkToolButtonGetType}
\alias{gtkToolItemGetType}
\alias{gtkToolbarGetType}
\alias{gtkTooltipsGetType}
\alias{gtkTreeDragDestGetType}
\alias{gtkTreeDragSourceGetType}
\alias{gtkTreeIterGetType}
\alias{gtkTreeModelFilterGetType}
\alias{gtkTreeModelGetType}
\alias{gtkTreeModelSortGetType}
\alias{gtkTreeRowReferenceGetType}
\alias{gtkTreeSelectionGetType}
\alias{gtkTreeSortableGetType}
\alias{gtkTreeStoreGetType}
\alias{gtkTreeViewColumnGetType}
\alias{gtkTreeViewGetType}
\alias{gtkTreeViewGridLinesGetType}
\alias{gtkUnitGetType} 
\alias{gtkUIManagerGetType}
\alias{gtkVBoxGetType}
\alias{gtkVButtonBoxGetType}
\alias{gtkVPanedGetType}
\alias{gtkVRulerGetType}
\alias{gtkVScaleGetType}
\alias{gtkVScrollbarGetType}
\alias{gtkVSeparatorGetType}
\alias{gtkViewportGetType}
\alias{gtkWidgetGetType}
\alias{gtkWindowGetType}
\alias{gtkWindowGroupGetType}
\alias{pangoAttrClassGetType}
\alias{pangoAttrListGetType}
\alias{pangoCairoFontMapGetType}
\alias{pangoColorGetType}
\alias{pangoFontFaceGetType}
\alias{pangoFontFamilyGetType}
\alias{pangoFontGetType}
\alias{pangoFontMetricsGetType}
\alias{pangoGlyphStringGetType}
\alias{pangoLayoutGetType}
\alias{pangoLayoutIterGetType}
\alias{pangoRendererGetType}
\alias{pangoTabArrayGetType}
\alias{gtkBuildableGetType}
\alias{gtkBuilderGetType}
\alias{gtkRecentActionGetType} 
\alias{gtkScaleButtonGetType}
\alias{gtkTooltipGetType}
\alias{gtkVolumeButtonGetType}
\alias{gAppInfoCreateFlagsGetType}
\alias{gAppInfoGetType}
\alias{gAppLaunchContextGetType}
\alias{gAskPasswordFlagsGetType}
\alias{gAsyncInitableGetType}
\alias{gAsyncResultGetType}
\alias{gBufferedInputStreamGetType}
\alias{gBufferedOutputStreamGetType}
\alias{gCancellableGetType}
\alias{gDataInputStreamGetType}
\alias{gDataOutputStreamGetType}
\alias{gDataStreamByteOrderGetType}
\alias{gDataStreamNewlineTypeGetType}
\alias{gDriveGetType}
\alias{gEmblemGetType}
\alias{gEmblemedIconGetType}
\alias{gFileAttributeInfoFlagsGetType}
\alias{gFileAttributeStatusGetType}
\alias{gFileAttributeTypeGetType}
\alias{gFileCopyFlagsGetType}
\alias{gFileCreateFlagsGetType}
\alias{gFileEnumeratorGetType}
\alias{gFileGetType}
\alias{gFileIOStreamGetType}
\alias{gFileIconGetType}
\alias{gFileInfoGetType}
\alias{gFileInputStreamGetType}
\alias{gFileMonitorEventGetType}
\alias{gFileMonitorFlagsGetType}
\alias{gFileMonitorGetType}
\alias{gFileOutputStreamGetType}
\alias{gFileQueryInfoFlagsGetType}
\alias{gFileTypeGetType}
\alias{gFilenameCompleterGetType}
\alias{gFilterInputStreamGetType}
\alias{gFilterOutputStreamGetType}
\alias{gIOModuleGetType}
\alias{gIOStreamGetType}
\alias{gIconGetType}
\alias{gInetAddressGetType}
\alias{gInitableGetType}
\alias{gInputStreamGetType}
\alias{gIoErrorEnumGetType}
\alias{gLoadableIconGetType}
\alias{gMemoryInputStreamGetType}
\alias{gMemoryOutputStreamGetType}
\alias{gMountGetType}
\alias{gMountOperationGetType}
\alias{gNativeVolumeMonitorGetType}
\alias{gNetworkAddressGetType}
\alias{gNetworkServiceGetType}
\alias{gOutputStreamGetType}
\alias{gOutputStreamSpliceFlagsGetType}
\alias{gPasswordSaveGetType}
\alias{gResolverGetType}
\alias{gSeekableGetType}
\alias{gSimpleAsyncResultGetType}
\alias{gSocketAddressEnumeratorGetType}
\alias{gSocketAddressGetType}
\alias{gSocketClientGetType}
\alias{gSocketConnectableGetType}
\alias{gSocketConnectionGetType}
\alias{gSocketControlMessageClassGetType}
\alias{gSocketControlMessageGetType}
\alias{gSocketGetType}
\alias{gSocketListenerGetType}
\alias{gSocketServiceGetType}
\alias{gSrvTargetGetType}
\alias{gTcpConnectionGetType}
\alias{gThemedIconGetType}
\alias{gThreadedSocketServiceGetType}
\alias{gVfsGetType}
\alias{gVolumeGetType}
\alias{gVolumeMonitorGetType}
\alias{gdkAppLaunchContextGetType}
\alias{gtkActivatableGetType}
\alias{gtkCellRendererSpinnerGetType}
\alias{gtkEntryBufferGetType}
\alias{gtkHSVGetType}
\alias{gtkInfoBarGetType}
\alias{gtkMountOperationGetType}
\alias{gtkOffscreenWindowGetType}
\alias{gtkOrientableGetType}
\alias{gtkSpinnerGetType}
\alias{gtkToolItemGroupGetType}
\alias{gtkToolPaletteGetType}
\name{typeMacros}
\title{GType accessor macros}
\description{This function returns the R representation of the GType of the object argument.}
\author{Michael Lawrence}
\keyword{interface}
\keyword{internal}
